B

Documentation

data X

Should show up on the page for both modules A and B

Constructors

X
Doc for consructor

data A

Constructors

A

test2 :: Bool

Doc for test2

reExport :: Int

Should show up on the page for both modules A and B

test :: Int

This link shouldn't work: other. These links should work: other, sortBy, test2, test2, fromMaybe. Module link: Prelude.

reExport :: Int

Should show up on the page for both modules A and B

data X

Should show up on the page for both modules A and B

Constructors

X
Doc for consructor