BinFingerprint

Description

Computing fingerprints of values serializeable with GHC's Binary module.

Computing fingerprints

fingerprintBinMem :: BinHandle -> IO Fingerprint

computeFingerprint :: Binary a => (BinHandle -> Name -> IO ()) -> a -> IO Fingerprint

putNameLiterally :: BinHandle -> Name -> IO ()

Used when we want to fingerprint a structure without depending on the fingerprints of external Names that it refers to.