ClsInst

Documentation

matchGlobalInst :: DynFlags -> Bool -> Class -> [Type] -> TcM ClsInstResult

data ClsInstResult

Constructors

NoInstance

OneInst

Fields

cir_new_theta :: [TcPredType]

cir_mk_ev :: [EvExpr] -> EvTerm

cir_what :: InstanceWhat

NotSure

data InstanceWhat

Constructors

BuiltinInstance

LocalInstance

TopLevInstance

Fields

iw_dfun_id :: DFunId

iw_safe_over :: SafeOverlapping

safeOverlap :: InstanceWhat -> Bool

data AssocInstInfo

Extra information about the parent instance declaration, needed when type-checking associated types. The Class is the enclosing class, the [TyVar] are the scoped type variable of the instance decl. The VarEnv Type maps class variables to their instance types.

Constructors

NotAssociated

InClsInst

Fields

ai_class :: Class

ai_tyvars :: [TyVar]

The scoped tyvars of the instance Why scoped? See bind_me in TcValidity.checkConsistentFamInst

ai_inst_env :: VarEnv Type

Maps class tyvars to their instance types See Note [Matching in the consistent-instantation check]

isNotAssociated :: AssocInstInfo -> Bool