CmmCallConv

Documentation

data ParamLocation

Constructors

RegisterParam GlobalReg

StackParam ByteOff

assignArgumentsPos :: DynFlags -> ByteOff -> Convention -> (a -> CmmType) -> [a] -> (ByteOff, [(a, ParamLocation)])

Given a list of arguments, and a function that tells their types, return a list showing where each argument is passed

assignStack :: DynFlags -> ByteOff -> (a -> CmmType) -> [a] -> (ByteOff, [(a, ParamLocation)])

realArgRegsCover :: DynFlags -> [GlobalReg]