Constraint

Documentation

data QCInst

Constructors

QCI

Fields

qci_ev :: CtEvidence

qci_tvs :: [TcTyVar]

qci_pred :: TcPredType

qci_pend_sc :: Bool

isPendingScInst :: QCInst -> Maybe QCInst

type Xi = Type

data Ct

Constructors

CDictCan

Fields

cc_ev :: CtEvidence

cc_class :: Class

cc_tyargs :: [Xi]

cc_pend_sc :: Bool

CIrredCan

Fields

cc_ev :: CtEvidence

cc_insol :: Bool

CTyEqCan

Fields

cc_ev :: CtEvidence

cc_tyvar :: TcTyVar

cc_rhs :: TcType

cc_eq_rel :: EqRel

CFunEqCan

Fields

cc_ev :: CtEvidence

cc_fun :: TyCon

cc_tyargs :: [Xi]

cc_fsk :: TcTyVar

CNonCanonical

Fields

cc_ev :: CtEvidence

CHoleCan

Fields

cc_ev :: CtEvidence

cc_hole :: Hole

CQuantCan QCInst

type Cts = Bag Ct

emptyCts :: Cts

andCts :: Cts -> Cts -> Cts

andManyCts :: [Cts] -> Cts

pprCts :: Cts -> SDoc

singleCt :: Ct -> Cts

listToCts :: [Ct] -> Cts

ctsElts :: Cts -> [Ct]

consCts :: Ct -> Cts -> Cts

snocCts :: Cts -> Ct -> Cts

extendCtsList :: Cts -> [Ct] -> Cts

isEmptyCts :: Cts -> Bool

isCTyEqCan :: Ct -> Bool

isCFunEqCan :: Ct -> Bool

isPendingScDict :: Ct -> Maybe Ct

superClassesMightHelp :: WantedConstraints -> Bool

True if taking superclasses of givens, or of wanteds (to perhaps expose more equalities or functional dependencies) might help to solve this constraint. See Note [When superclasses help]

getPendingWantedScs :: Cts -> ([Ct], Cts)

isCDictCan_Maybe :: Ct -> Maybe Class

isCFunEqCan_maybe :: Ct -> Maybe (TyCon, [Type])

isCNonCanonical :: Ct -> Bool

isWantedCt :: Ct -> Bool

isDerivedCt :: Ct -> Bool

isGivenCt :: Ct -> Bool

isHoleCt :: Ct -> Bool

isOutOfScopeCt :: Ct -> Bool

isExprHoleCt :: Ct -> Bool

isTypeHoleCt :: Ct -> Bool

isUserTypeErrorCt :: Ct -> Bool

getUserTypeErrorMsg :: Ct -> Maybe Type

A constraint is considered to be a custom type error, if it contains custom type errors anywhere in it. See Note [Custom type errors in constraints]

ctEvidence :: Ct -> CtEvidence

ctLoc :: Ct -> CtLoc

setCtLoc :: Ct -> CtLoc -> Ct

ctPred :: Ct -> PredType

ctFlavour :: Ct -> CtFlavour

Get the flavour of the given Ct

ctEqRel :: Ct -> EqRel

Get the equality relation for the given Ct

ctOrigin :: Ct -> CtOrigin

ctEvId :: Ct -> EvVar

mkTcEqPredLikeEv :: CtEvidence -> TcType -> TcType -> TcType

Makes a new equality predicate with the same role as the given evidence.

mkNonCanonical :: CtEvidence -> Ct

mkNonCanonicalCt :: Ct -> Ct

mkGivens :: CtLoc -> [EvId] -> [Ct]

mkIrredCt :: CtEvidence -> Ct

mkInsolubleCt :: CtEvidence -> Ct

ctEvPred :: CtEvidence -> TcPredType

ctEvLoc :: CtEvidence -> CtLoc

ctEvOrigin :: CtEvidence -> CtOrigin

ctEvEqRel :: CtEvidence -> EqRel

Get the equality relation relevant for a CtEvidence

ctEvExpr :: CtEvidence -> EvExpr

ctEvTerm :: CtEvidence -> EvTerm

ctEvCoercion :: HasDebugCallStack => CtEvidence -> TcCoercion

ctEvEvId :: CtEvidence -> EvVar

tyCoVarsOfCt :: Ct -> TcTyCoVarSet

Returns free variables of constraints as a non-deterministic set

tyCoVarsOfCts :: Cts -> TcTyCoVarSet

Returns free variables of a bag of constraints as a non-deterministic set. See Note [Deterministic FV] in FV.

tyCoVarsOfCtList :: Ct -> [TcTyCoVar]

Returns free variables of constraints as a deterministically ordered. list. See Note [Deterministic FV] in FV.

tyCoVarsOfCtsList :: Cts -> [TcTyCoVar]

Returns free variables of a bag of constraints as a deterministically odered list. See Note [Deterministic FV] in FV.

data WantedConstraints

Constructors

WC

Fields

wc_simple :: Cts

wc_impl :: Bag Implication

insolubleWC :: WantedConstraints -> Bool

emptyWC :: WantedConstraints

isEmptyWC :: WantedConstraints -> Bool

isSolvedWC :: WantedConstraints -> Bool

Checks whether a the given wanted constraints are solved, i.e. that there are no simple constraints left and all the implications are solved.

andWC :: WantedConstraints -> WantedConstraints -> WantedConstraints

unionsWC :: [WantedConstraints] -> WantedConstraints

mkSimpleWC :: [CtEvidence] -> WantedConstraints

mkImplicWC :: Bag Implication -> WantedConstraints

addInsols :: WantedConstraints -> Bag Ct -> WantedConstraints

insolublesOnly :: WantedConstraints -> WantedConstraints

addSimples :: WantedConstraints -> Bag Ct -> WantedConstraints

addImplics :: WantedConstraints -> Bag Implication -> WantedConstraints

tyCoVarsOfWC :: WantedConstraints -> TyCoVarSet

Returns free variables of WantedConstraints as a non-deterministic set. See Note [Deterministic FV] in FV.

dropDerivedWC :: WantedConstraints -> WantedConstraints

dropDerivedSimples :: Cts -> Cts

tyCoVarsOfWCList :: WantedConstraints -> [TyCoVar]

Returns free variables of WantedConstraints as a deterministically ordered list. See Note [Deterministic FV] in FV.

insolubleCt :: Ct -> Bool

insolubleEqCt :: Ct -> Bool

isDroppableCt :: Ct -> Bool

insolubleImplic :: Implication -> Bool

arisesFromGivens :: Ct -> Bool

data Implication

Constructors

Implic

Fields

ic_tclvl :: TcLevel

ic_skols :: [TcTyVar]

ic_info :: SkolemInfo

ic_telescope :: Maybe SDoc

ic_given :: [EvVar]

ic_no_eqs :: Bool

ic_warn_inaccessible :: Bool

ic_env :: TcLclEnv

ic_wanted :: WantedConstraints

ic_binds :: EvBindsVar

ic_need_inner :: VarSet

ic_need_outer :: VarSet

ic_status :: ImplicStatus

implicationPrototype :: Implication

data ImplicStatus

Constructors

IC_Solved

Fields

ics_dead :: [EvVar]

IC_Insoluble

IC_BadTelescope

IC_Unsolved

isInsolubleStatus :: ImplicStatus -> Bool

isSolvedStatus :: ImplicStatus -> Bool

data SubGoalDepth

See Note [SubGoalDepth]

initialSubGoalDepth :: SubGoalDepth

maxSubGoalDepth :: SubGoalDepth -> SubGoalDepth -> SubGoalDepth

bumpSubGoalDepth :: SubGoalDepth -> SubGoalDepth

subGoalDepthExceeded :: DynFlags -> SubGoalDepth -> Bool

data CtLoc

Constructors

CtLoc

Fields

ctl_origin :: CtOrigin

ctl_env :: TcLclEnv

ctl_t_or_k :: Maybe TypeOrKind

ctl_depth :: !SubGoalDepth

ctLocSpan :: CtLoc -> RealSrcSpan

ctLocEnv :: CtLoc -> TcLclEnv

ctLocLevel :: CtLoc -> TcLevel

ctLocOrigin :: CtLoc -> CtOrigin

ctLocTypeOrKind_maybe :: CtLoc -> Maybe TypeOrKind

ctLocDepth :: CtLoc -> SubGoalDepth

bumpCtLocDepth :: CtLoc -> CtLoc

isGivenLoc :: CtLoc -> Bool

setCtLocOrigin :: CtLoc -> CtOrigin -> CtLoc

updateCtLocOrigin :: CtLoc -> (CtOrigin -> CtOrigin) -> CtLoc

setCtLocEnv :: CtLoc -> TcLclEnv -> CtLoc

setCtLocSpan :: CtLoc -> RealSrcSpan -> CtLoc

pprCtLoc :: CtLoc -> SDoc

data CtEvidence

Constructors

CtGiven

Fields

ctev_pred :: TcPredType

ctev_evar :: EvVar

ctev_loc :: CtLoc

CtWanted

Fields

ctev_pred :: TcPredType

ctev_dest :: TcEvDest

ctev_nosh :: ShadowInfo

ctev_loc :: CtLoc

CtDerived

Fields

ctev_pred :: TcPredType

ctev_loc :: CtLoc

data TcEvDest

A place for type-checking evidence to go after it is generated. Wanted equalities are always HoleDest; other wanteds are always EvVarDest.

Constructors

EvVarDest EvVar
bind this var to the evidence EvVarDest is always used for non-type-equalities e.g. class constraints

HoleDest CoercionHole
fill in this hole with the evidence HoleDest is always used for type-equalities See Note [Coercion holes] in TyCoRep

mkKindLoc :: TcType -> TcType -> CtLoc -> CtLoc

toKindLoc :: CtLoc -> CtLoc

Take a CtLoc and moves it to the kind level

mkGivenLoc :: TcLevel -> SkolemInfo -> TcLclEnv -> CtLoc

isWanted :: CtEvidence -> Bool

isGiven :: CtEvidence -> Bool

isDerived :: CtEvidence -> Bool

isGivenOrWDeriv :: CtFlavour -> Bool

ctEvRole :: CtEvidence -> Role

Get the role relevant for a CtEvidence

wrapType :: Type -> [TyVar] -> [PredType] -> Type

wrapTypeWithImplication :: Type -> Implication -> Type

Wraps the given type with the constraints (via ic_given) in the given implication, according to the variables mentioned (via ic_skols) in the implication, but taking care to only wrap those variables that are mentioned in the type or the implication.

data CtFlavour

Constructors

Given

Wanted ShadowInfo

Derived

data ShadowInfo

Constructors

WDeriv

WOnly

ctEvFlavour :: CtEvidence -> CtFlavour

type CtFlavourRole = (CtFlavour, EqRel)

Whether or not one Ct can rewrite another is determined by its flavour and its equality relation. See also Note [Flavours with roles] in TcSMonad

ctEvFlavourRole :: CtEvidence -> CtFlavourRole

Extract the flavour, role, and boxity from a CtEvidence

ctFlavourRole :: Ct -> CtFlavourRole

Extract the flavour and role from a Ct

eqCanRewrite :: EqRel -> EqRel -> Bool

eqCanRewriteFR :: CtFlavourRole -> CtFlavourRole -> Bool

eqMayRewriteFR :: CtFlavourRole -> CtFlavourRole -> Bool

eqCanDischargeFR :: CtFlavourRole -> CtFlavourRole -> Bool

funEqCanDischarge :: CtEvidence -> CtEvidence -> (SwapFlag, Bool)

funEqCanDischargeF :: CtFlavour -> CtFlavour -> (SwapFlag, Bool)

pprEvVarTheta :: [EvVar] -> SDoc

pprEvVars :: [EvVar] -> SDoc

pprEvVarWithType :: EvVar -> SDoc

data Hole

An expression or type hole

Constructors

ExprHole UnboundVar
Either an out-of-scope variable or a "true" hole in an expression (TypedHoles)

TypeHole OccName
A hole in a type (PartialTypeSignatures)

holeOcc :: Hole -> OccName