CostCentre

Documentation

data CostCentre

A Cost Centre is a single {--} annotation.

Constructors

NormalCC

Fields

cc_flavour :: CCFlavour

Two cost centres may have the same name and module but different SrcSpans, so we need a way to distinguish them easily and give them different object-code labels. So every CostCentre has an associated flavour that indicates how it was generated, and flavours that allow multiple instances of the same name and module have a deterministic 0-based index.

cc_name :: CcName

Name of the cost centre itself

cc_mod :: Module

Name of module defining this CC.

cc_loc :: SrcSpan

AllCafsCC

Fields

cc_mod :: Module

Name of module defining this CC.

cc_loc :: SrcSpan

type CcName = FastString

data CCFlavour

The flavour of a cost centre.

Index fields represent 0-based indices giving source-code ordering of centres with the same module, name, and flavour.

Constructors

CafCC
Auto-generated top-level thunk

ExprCC !CostCentreIndex
Explicitly annotated expression

DeclCC !CostCentreIndex
Explicitly annotated declaration

HpcCC !CostCentreIndex
Generated by HPC for coverage

data CostCentreStack

A Cost Centre Stack is something that can be attached to a closure. This is either:

the current cost centre stack (CCCS)

a pre-defined cost centre stack (there are several pre-defined CCSs, see below).

type CollectedCCs = ([CostCentre], [CostCentreStack])

emptyCollectedCCs :: CollectedCCs

collectCC :: CostCentre -> CostCentreStack -> CollectedCCs -> CollectedCCs

currentCCS :: CostCentreStack

dontCareCCS :: CostCentreStack

isCurrentCCS :: CostCentreStack -> Bool

maybeSingletonCCS :: CostCentreStack -> Maybe CostCentre

mkUserCC :: FastString -> Module -> SrcSpan -> CCFlavour -> CostCentre

mkAutoCC :: Id -> Module -> CostCentre

mkAllCafsCC :: Module -> SrcSpan -> CostCentre

mkSingletonCCS :: CostCentre -> CostCentreStack

isCafCCS :: CostCentreStack -> Bool

isCafCC :: CostCentre -> Bool

isSccCountCC :: CostCentre -> Bool

Is this a cost-centre which records scc counts

sccAbleCC :: CostCentre -> Bool

Is this a cost-centre which can be sccd ?

ccFromThisModule :: CostCentre -> Module -> Bool

pprCostCentreCore :: CostCentre -> SDoc

Print a flavour in Core

costCentreUserName :: CostCentre -> String

costCentreUserNameFS :: CostCentre -> FastString

costCentreSrcSpan :: CostCentre -> SrcSpan

cmpCostCentre :: CostCentre -> CostCentre -> Ordering