DataCon

Main data types

data DataCon

A data constructor

AnnKeywordId : AnnOpen, AnnClose,AnnComma

data DataConRep

Data Constructor Representation

Constructors

NoDataConRep

DCR

Fields

dcr_wrap_id :: Id

dcr_boxer :: DataConBoxer

dcr_arg_tys :: [Type]

dcr_stricts :: [StrictnessMark]

dcr_bangs :: [HsImplBang]

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie '~'

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{--} specified

SrcNoUnpack
{--} specified

NoSrcUnpack
no unpack pragma

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as the user wrote them in the source code.

(HsSrcBang _ SrcUnpack SrcLazy) and (HsSrcBang _ SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsSrcBang _ NoSrcUnpack SrcLazy)

Constructors

HsSrcBang SourceText SrcUnpackedness SrcStrictness

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field

HsStrict
Strict but not unpacked field

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data StrictnessMark

Constructors

MarkedStrict

NotMarkedStrict

type ConTag = Int

Constructor Tag

Type of the tags associated with each constructor possibility or superclass selector

Equality specs

data EqSpec

An EqSpec is a tyvar/type pair representing an equality made in rejigging a GADT constructor

mkEqSpec :: TyVar -> Type -> EqSpec

Make an EqSpec

eqSpecTyVar :: EqSpec -> TyVar

eqSpecType :: EqSpec -> Type

eqSpecPair :: EqSpec -> (TyVar, Type)

eqSpecPreds :: [EqSpec] -> ThetaType

substEqSpec :: TCvSubst -> EqSpec -> EqSpec

Substitute in an EqSpec. Precondition: if the LHS of the EqSpec is mapped in the substitution, it is mapped to a type variable, not a full type.

filterEqSpec :: [EqSpec] -> [TyVarBinder] -> [TyVarBinder]

Filter out any TyBinders mentioned in an EqSpec

Field labels

data FieldLbl a

Fields in an algebraic record type

Constructors

FieldLabel

Fields

flLabel :: FieldLabelString

User-visible label of the field

flIsOverloaded :: Bool

Was DuplicateRecordFields on in the defining module for this datatype?

flSelector :: a

Record selector function

type FieldLabel = FieldLbl Name

type FieldLabelString = FastString

Field labels are just represented as strings; they are not necessarily unique (even within a module)

Type construction

mkDataCon

Arguments

:: Name

-> Bool
Is the constructor declared infix?

-> TyConRepName
TyConRepName for the promoted TyCon

-> [HsSrcBang]
Strictness/unpack annotations, from user

-> [FieldLabel]
Field labels for the constructor, if it is a record, otherwise empty

-> [TyVarBinder]
Universals. See Note [TyVarBinders in DataCons]

-> [TyVarBinder]
Existentials. (These last two must be Named and Inferred/Specified)

-> [EqSpec]
GADT equalities

-> ThetaType
Theta-type occuring before the arguments proper

-> [Type]
Original argument types

-> Type
Original result type

-> RuntimeRepInfo
See comments on RuntimeRepInfo

-> TyCon
Representation type constructor

-> ThetaType
The "stupid theta", context of the data declaration e.g. data Eq a => T a ...

-> Id
Worker Id

-> DataConRep
Representation

-> DataCon

Build a new data constructor

buildAlgTyCon

Arguments

:: Name

-> [TyVar]
Kind variables and type variables

-> [Role]

-> Maybe CType

-> ThetaType
Stupid theta

-> AlgTyConRhs

-> Bool
True = was declared in GADT syntax

-> AlgTyConFlav

-> TyCon

buildSynTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind

-> [Role]

-> Type

-> TyCon

fIRST_TAG :: ConTag

Tags are allocated from here for real constructors or for superclass selectors

Type deconstruction

dataConRepType :: DataCon -> Type

The representation type of the data constructor, i.e. the sort type that will represent values of this type at runtime

dataConSig :: DataCon -> ([TyVar], ThetaType, [Type], Type)

The "signature" of the DataCon returns, in order:

1) The result of dataConAllTyVars,

2) All the ThetaTypes relating to the DataCon (coercion, dictionary, implicit parameter - whatever)

3) The type arguments to the constructor

4) The original result type of the DataCon

dataConInstSig :: DataCon -> [Type] -> ([TyVar], ThetaType, [Type])

Instantantiate the universal tyvars of a data con, returning the instantiated existentials, constraints, and args

dataConFullSig :: DataCon -> ([TyVar], [TyVar], [EqSpec], ThetaType, [Type], Type)

The "full signature" of the DataCon returns, in order:

1) The result of dataConUnivTyVars

2) The result of dataConExTyVars

3) The GADT equalities

4) The result of dataConDictTheta

5) The original argument types to the DataCon (i.e. before any change of the representation of the type)

6) The original result type of the DataCon

dataConName :: DataCon -> Name

The Name of the DataCon, giving it a unique, rooted identification

dataConIdentity :: DataCon -> [Word8]

The string package:module.name identifying a constructor, which is attached to its info table and used by the GHCi debugger and the heap profiler

dataConTag :: DataCon -> ConTag

The tag used for ordering DataCons

dataConTyCon :: DataCon -> TyCon

The type constructor that we are building via this data constructor

dataConOrigTyCon :: DataCon -> TyCon

The original type constructor used in the definition of this data constructor. In case of a data family instance, that will be the family type constructor.

dataConUserType :: DataCon -> Type

The user-declared type of the data constructor in the nice-to-read form:

T :: forall a b. a -> b -> T [a]

rather than:

T :: forall a c. forall b. (c~[a]) => a -> b -> T c

NB: If the constructor is part of a data instance, the result type mentions the family tycon, not the internal one.

dataConUnivTyVars :: DataCon -> [TyVar]

The universally-quantified type variables of the constructor

dataConUnivTyVarBinders :: DataCon -> [TyVarBinder]

TyBinders for the universally-quantified type variables

dataConExTyVars :: DataCon -> [TyVar]

The existentially-quantified type variables of the constructor

dataConExTyVarBinders :: DataCon -> [TyVarBinder]

TyBinders for the existentially-quantified type variables

dataConAllTyVars :: DataCon -> [TyVar]

Both the universal and existentiatial type variables of the constructor

dataConEqSpec :: DataCon -> [EqSpec]

Equalities derived from the result type of the data constructor, as written by the programmer in any GADT declaration. This includes *all* GADT-like equalities, including those written in by hand by the programmer.

dataConTheta :: DataCon -> ThetaType

The *full* constraints on the constructor type.

dataConStupidTheta :: DataCon -> ThetaType

The "stupid theta" of the DataCon, such as data Eq a in:

data Eq a => T a = ...

dataConInstArgTys

Arguments

:: DataCon
A datacon with no existentials or equality constraints However, it can have a dcTheta (notably it can be a class dictionary, with superclasses)

-> [Type]
Instantiated at these types

-> [Type]

Finds the instantiated types of the arguments required to construct a DataCon representation NB: these INCLUDE any dictionary args but EXCLUDE the data-declaration context, which is discarded It's all post-flattening etc; this is a representation type

dataConOrigArgTys :: DataCon -> [Type]

Returns the argument types of the wrapper, excluding all dictionary arguments and without substituting for any type variables

dataConOrigResTy :: DataCon -> Type

dataConInstOrigArgTys :: DataCon -> [Type] -> [Type]

Returns just the instantiated value argument types of a DataCon, (excluding dictionary args)

dataConRepArgTys :: DataCon -> [Type]

Returns the arg types of the worker, including *all* evidence, after any flattening has been done and without substituting for any type variables

dataConFieldLabels :: DataCon -> [FieldLabel]

The labels for the fields of this particular DataCon

dataConFieldType :: DataCon -> FieldLabelString -> Type

Extract the type for any given labelled field of the DataCon

dataConFieldType_maybe :: DataCon -> FieldLabelString -> Maybe (FieldLabel, Type)

Extract the label and type for any given labelled field of the DataCon, or return Nothing if the field does not belong to it

dataConSrcBangs :: DataCon -> [HsSrcBang]

Strictness/unpack annotations, from user; or, for imported DataCons, from the interface file The list is in one-to-one correspondence with the arity of the DataCon

dataConSourceArity :: DataCon -> Arity

Source-level arity of the data constructor

dataConRepArity :: DataCon -> Arity

Gives the number of actual fields in the representation of the data constructor. This may be more than appear in the source code; the extra ones are the existentially quantified dictionaries

dataConIsInfix :: DataCon -> Bool

Should the DataCon be presented infix?

dataConWorkId :: DataCon -> Id

Get the Id of the DataCon worker: a function that is the "actual" constructor and has no top level binding in the program. The type may be different from the obvious one written in the source program. Panics if there is no such Id for this DataCon

dataConWrapId :: DataCon -> Id

Returns an Id which looks like the Haskell-source constructor by using the wrapper if it exists (see dataConWrapId_maybe) and failing over to the worker (see dataConWorkId)

dataConWrapId_maybe :: DataCon -> Maybe Id

Get the Id of the DataCon wrapper: a function that wraps the "actual" constructor so it has the type visible in the source program: c.f. dataConWorkId. Returns Nothing if there is no wrapper, which occurs for an algebraic data constructor and also for a newtype (whose constructor is inlined compulsorily)

dataConImplicitTyThings :: DataCon -> [TyThing]

Find all the Ids implicitly brought into scope by the data constructor. Currently, the union of the dataConWorkId and the dataConWrapId

dataConRepStrictness :: DataCon -> [StrictnessMark]

Give the demands on the arguments of a Core constructor application (Con dc args)

dataConImplBangs :: DataCon -> [HsImplBang]

dataConBoxer :: DataCon -> Maybe DataConBoxer

splitDataProductType_maybe

Arguments

:: Type
A product type, perhaps

-> Maybe (TyCon, [Type], DataCon, [Type])

Extract the type constructor, type argument, data constructor and it's representation argument types from a type if it is a product type.

Precisely, we return Just for any type that is all of:

Concrete (i.e. constructors visible)

Single-constructor

Not existentially quantified

Whether the type is a data type or a newtype

Predicates on DataCons

isNullarySrcDataCon :: DataCon -> Bool

Return whether there are any argument types for this DataCons original source type See Note [DataCon arities]

isNullaryRepDataCon :: DataCon -> Bool

Return whether there are any argument types for this DataCons runtime representation type See Note [DataCon arities]

isTupleDataCon :: DataCon -> Bool

isUnboxedTupleCon :: DataCon -> Bool

isUnboxedSumCon :: DataCon -> Bool

isVanillaDataCon :: DataCon -> Bool

Vanilla DataCons are those that are nice boring Haskell 98 constructors

classDataCon :: Class -> DataCon

dataConCannotMatch :: [Type] -> DataCon -> Bool

isBanged :: HsImplBang -> Bool

isMarkedStrict :: StrictnessMark -> Bool

eqHsBang :: HsImplBang -> HsImplBang -> Bool

Compare strictness annotations

isSrcStrict :: SrcStrictness -> Bool

isSrcUnpacked :: SrcUnpackedness -> Bool

specialPromotedDc :: DataCon -> Bool

Should this DataCon be allowed in a type even without -XDataKinds? Currently, only Lifted & Unlifted

isLegacyPromotableDataCon :: DataCon -> Bool

Was this datacon promotable before GHC 8.0? That is, is it promotable without -XTypeInType

isLegacyPromotableTyCon :: TyCon -> Bool

Was this tycon promotable before GHC 8.0? That is, is it promotable without -XTypeInType

Promotion related functions

promoteDataCon :: DataCon -> TyCon