DebuggerUtils

Documentation

dataConInfoPtrToName :: Ptr () -> TcM (Either String Name)

Given a data constructor in the heap, find its Name. The info tables for data constructors have a field which records the source name of the constructor as a Ptr Word8 (UTF-8 encoded string). The format is:

Package:Module.Name

We use this string to lookup the interpreter's internal representation of the name using the lookupOrig.