Desugar

Desugaring operations

deSugar :: HscEnv -> ModLocation -> TcGblEnv -> IO (Messages, Maybe ModGuts)

Main entry point to the desugarer.

deSugarExpr :: HscEnv -> LHsExpr Id -> IO (Messages, Maybe CoreExpr)

Dependency/fingerprinting code (used by MkIface)

mkUsageInfo :: HscEnv -> Module -> ImportedMods -> NameSet -> [FilePath] -> IO [Usage]

mkUsedNames :: TcGblEnv -> NameSet

mkDependencies :: TcGblEnv -> IO Dependencies

Extract information from the rename and typecheck phases to produce a dependencies information for the module being compiled.