Digraph

Documentation

data Graph node

graphFromEdgedVerticesOrd :: Ord key => [Node key payload] -> Graph (Node key payload)

graphFromEdgedVerticesUniq :: Uniquable key => [Node key payload] -> Graph (Node key payload)

data SCC vertex

Strongly connected component.

Constructors

AcyclicSCC vertex
A single vertex that is not in any cycle.

CyclicSCC [vertex]
A maximal set of mutually reachable vertices.

data Node key payload

Constructors

DigraphNode

Fields

node_payload :: payload

node_key :: key

node_dependencies :: [key]

flattenSCC :: SCC vertex -> [vertex]

The vertices of a strongly connected component.

flattenSCCs :: [SCC a] -> [a]

The vertices of a list of strongly connected components.

stronglyConnCompG :: Graph node -> [SCC node]

topologicalSortG :: Graph node -> [node]

verticesG :: Graph node -> [node]

edgesG :: Graph node -> [Edge node]

hasVertexG :: Graph node -> node -> Bool

reachableG :: Graph node -> node -> [node]

reachablesG :: Graph node -> [node] -> [node]

transposeG :: Graph node -> Graph node

emptyG :: Graph node -> Bool

findCycle :: forall payload key. Ord key => [Node key payload] -> Maybe [payload]

Find a reasonably short cycle a->b->c->a, in a strongly connected component. The input nodes are presumed to be a SCC, so you can start anywhere.

stronglyConnCompFromEdgedVerticesOrd :: Ord key => [Node key payload] -> [SCC payload]

stronglyConnCompFromEdgedVerticesOrdR :: Ord key => [Node key payload] -> [SCC (Node key payload)]

stronglyConnCompFromEdgedVerticesUniq :: Uniquable key => [Node key payload] -> [SCC payload]

stronglyConnCompFromEdgedVerticesUniqR :: Uniquable key => [Node key payload] -> [SCC (Node key payload)]