DsUsage

Dependency/fingerprinting code (used by MkIface)

mkUsageInfo :: HscEnv -> Module -> ImportedMods -> NameSet -> [FilePath] -> [(Module, Fingerprint)] -> [ModIface] -> IO [Usage]

mkUsedNames :: TcGblEnv -> NameSet

mkDependencies :: InstalledUnitId -> [Module] -> TcGblEnv -> IO Dependencies

Extract information from the rename and typecheck phases to produce a dependencies information for the module being compiled.

The first argument is additional dependencies from plugins