Format

Description

Formats on this architecture A Format is a combination of width and class

TODO: Signed vs unsigned?

TODO: This module is currenly shared by all architectures because NCGMonad need to know about it to make a VReg. It would be better to have architecture specific formats, and do the overloading properly. eg SPARC doesn't care about FF80.

Documentation

data Format

Constructors

II8

II16

II32

II64

FF32

FF64

FF80

intFormat :: Width -> Format

Get the integer format of this width.

floatFormat :: Width -> Format

Get the float format of this width.

isFloatFormat :: Format -> Bool

Check if a format represents a floating point value.

cmmTypeFormat :: CmmType -> Format

Convert a Cmm type to a Format.

formatToWidth :: Format -> Width

Get the Width of a Format.

formatInBytes :: Format -> Int