GHC

Initialisation

defaultErrorHandler :: (ExceptionMonad m, MonadIO m) => LogAction -> m a -> m a

Install some default exception handlers and run the inner computation. Unless you want to handle exceptions yourself, you should wrap this around the top level of your program. The default handlers output the error message(s) to stderr and exit cleanly.

defaultCleanupHandler :: (ExceptionMonad m, MonadIO m) => DynFlags -> m a -> m a

Install a default cleanup handler to remove temporary files deposited by a GHC run. This is seperate from defaultErrorHandler, because you might want to override the error handling, but still get the ordinary cleanup behaviour.

GHC Monad

data Ghc a

A minimal implementation of a GhcMonad. If you need a custom monad, e.g., to maintain additional state consider wrapping this monad or using GhcT.

data GhcT m a

A monad transformer to add GHC specific features to another monad.

Note that the wrapped monad must support IO and handling of exceptions.

class (Functor m, MonadIO m, ExceptionMonad m) => GhcMonad m where

A monad that has all the features needed by GHC API calls.

In short, a GHC monad

allows embedding of IO actions,

can log warnings,

allows handling of (extensible) exceptions, and

maintains a current session.

If you do not use Ghc or GhcT, make sure to call initGhcMonad before any call to the GHC API functions can occur.

Methods

getSession :: m HscEnv

setSession :: HscEnv -> m ()

data HscEnv

Hscenv is like Session, except that some of the fields are immutable. An HscEnv is used to compile a single module from plain Haskell source code (after preprocessing) to either C, assembly or C--. Things like the module graph don't change during a single compilation.

Historical note: "hsc" used to be the name of the compiler binary, when there was a separate driver and compiler. To compile a single module, the driver would invoke hsc on the source code... so nowadays we think of hsc as the layer of the compiler that deals with compiling a single module.

runGhc

Arguments

:: Maybe FilePath
See argument to initGhcMonad.

-> Ghc a
The action to perform.

-> IO a

Run function for the Ghc monad.

It initialises the GHC session and warnings via initGhcMonad. Each call to this function will create a new session which should not be shared among several threads.

Any errors not handled inside the Ghc action are propagated as IO exceptions.

runGhcT

Arguments

:: (ExceptionMonad m, Functor m, MonadIO m)

=> Maybe FilePath
See argument to initGhcMonad.

-> GhcT m a
The action to perform.

-> m a

Run function for GhcT monad transformer.

It initialises the GHC session and warnings via initGhcMonad. Each call to this function will create a new session which should not be shared among several threads.

initGhcMonad :: GhcMonad m => Maybe FilePath -> m ()

Initialise a GHC session.

If you implement a custom GhcMonad you must call this function in the monad run function. It will initialise the session variable and clear all warnings.

The first argument should point to the directory where GHC's library files reside. More precisely, this should be the output of ghc --print-libdir of the version of GHC the module using this API is compiled with. For portability, you should use the ghc-paths package, available at http://hackage.haskell.org/cgi-bin/hackage-scripts/package/ghc-paths.

gcatch :: (ExceptionMonad m, Exception e) => m a -> (e -> m a) -> m a

Generalised version of catch, allowing an arbitrary exception handling monad instead of just IO.

gbracket :: ExceptionMonad m => m a -> (a -> m b) -> (a -> m c) -> m c

Generalised version of bracket, allowing an arbitrary exception handling monad instead of just IO.

gfinally :: ExceptionMonad m => m a -> m b -> m a

Generalised version of finally, allowing an arbitrary exception handling monad instead of just IO.

printException :: GhcMonad m => SourceError -> m ()

Print the error message and all warnings. Useful inside exception handlers. Clears warnings after printing.

printExceptionAndWarnings :: GhcMonad m => SourceError -> m ()

handleSourceError

Arguments

:: ExceptionMonad m

=> (SourceError -> m a)
exception handler

-> m a
action to perform

-> m a

Perform the given action and call the exception handler if the action throws a SourceError. See SourceError for more information.

needsTemplateHaskell :: ModuleGraph -> Bool

Determines whether a set of modules requires Template Haskell.

Note that if the session's DynFlags enabled Template Haskell when depanal was called, then each module in the returned module graph will have Template Haskell enabled whether it is actually needed or not.

Flags and settings

data DynFlags

Contains not only a collection of DynFlags but also a plethora of information relating to the compilation of a single file or GHC session

Constructors

DynFlags

Fields

ghcMode :: GhcMode

ghcLink :: GhcLink

hscTarget :: HscTarget

settings :: Settings

hscOutName :: String

Name of the output file

extCoreName :: String

Name of the .hcr output file

verbosity :: Int

Verbosity level: see Note [Verbosity levels]

optLevel :: Int

Optimisation level

simplPhases :: Int

Number of simplifier phases

maxSimplIterations :: Int

Max simplifier iterations

shouldDumpSimplPhase :: Maybe String

ruleCheck :: Maybe String

strictnessBefore :: [Int]

Additional demand analysis

simplTickFactor :: Int

Multiplier for simplifier ticks

specConstrThreshold :: Maybe Int

Threshold for SpecConstr

specConstrCount :: Maybe Int

Max number of specialisations for any one function

liberateCaseThreshold :: Maybe Int

Threshold for LiberateCase

floatLamArgs :: Maybe Int

Arg count for lambda floating See CoreMonad.FloatOutSwitches

cmdlineHcIncludes :: [String]

-#includes

importPaths :: [FilePath]

mainModIs :: Module

mainFunIs :: Maybe String

ctxtStkDepth :: Int

Typechecker context stack depth

thisPackage :: PackageId

name of package currently being compiled

ways :: [Way]

Way flags from the command line

buildTag :: String

The global "way" (e.g. "p" for prof)

rtsBuildTag :: String

The RTS "way"

splitInfo :: Maybe (String, Int)

objectDir :: Maybe String

dylibInstallName :: Maybe String

hiDir :: Maybe String

stubDir :: Maybe String

dumpDir :: Maybe String

objectSuf :: String

hcSuf :: String

hiSuf :: String

outputFile :: Maybe String

outputHi :: Maybe String

dynLibLoader :: DynLibLoader

dumpPrefix :: Maybe FilePath

This is set by runPipeline based on where its output is going.

dumpPrefixForce :: Maybe FilePath

Override the dumpPrefix set by runPipeline. Set by -ddump-file-prefix

includePaths :: [String]

libraryPaths :: [String]

frameworkPaths :: [String]

cmdlineFrameworks :: [String]

rtsOpts :: Maybe String

rtsOptsEnabled :: RtsOptsEnabled

hpcDir :: String

Path to store the .mix files

pluginModNames :: [ModuleName]

pluginModNameOpts :: [(ModuleName, String)]

depMakefile :: FilePath

depIncludePkgDeps :: Bool

depExcludeMods :: [ModuleName]

depSuffixes :: [String]

extraPkgConfs :: [FilePath]

The -package-conf flags given on the command line, in the order they appeared.

packageFlags :: [PackageFlag]

The -package and -hide-package flags from the command-line

pkgDatabase :: Maybe [PackageConfig]

pkgState :: PackageState

filesToClean :: IORef [FilePath]

dirsToClean :: IORef (Map FilePath FilePath)

generatedDumps :: IORef (Set FilePath)

flags :: IntSet

warningFlags :: IntSet

language :: Maybe Language

safeHaskell :: SafeHaskellMode

Safe Haskell mode

thOnLoc :: SrcSpan

newDerivOnLoc :: SrcSpan

pkgTrustOnLoc :: SrcSpan

warnSafeOnLoc :: SrcSpan

warnUnsafeOnLoc :: SrcSpan

extensions :: [OnOff ExtensionFlag]

extensionFlags :: IntSet

log_action :: LogAction

Message output action: use ErrUtils instead of this if you can

haddockOptions :: Maybe String

profAuto :: ProfAuto

what kind of {--} to add automatically

data DynFlag

Enumerates the simple on-or-off dynamic flags

Constructors

Opt_D_dump_cmm

Opt_D_dump_raw_cmm

Opt_D_dump_cmmz

Opt_D_dump_cmmz_pretty

Opt_D_dump_cmmz_cbe

Opt_D_dump_cmmz_proc

Opt_D_dump_cmmz_spills

Opt_D_dump_cmmz_rewrite

Opt_D_dump_cmmz_dead

Opt_D_dump_cmmz_stub

Opt_D_dump_cmmz_sp

Opt_D_dump_cmmz_procmap

Opt_D_dump_cmmz_split

Opt_D_dump_cmmz_lower

Opt_D_dump_cmmz_info

Opt_D_dump_cmmz_cafs

Opt_D_dump_cps_cmm

Opt_D_dump_cvt_cmm

Opt_D_dump_asm

Opt_D_dump_asm_native

Opt_D_dump_asm_liveness

Opt_D_dump_asm_coalesce

Opt_D_dump_asm_regalloc

Opt_D_dump_asm_regalloc_stages

Opt_D_dump_asm_conflicts

Opt_D_dump_asm_stats

Opt_D_dump_asm_expanded

Opt_D_dump_llvm

Opt_D_dump_core_stats

Opt_D_dump_cpranal

Opt_D_dump_deriv

Opt_D_dump_ds

Opt_D_dump_flatC

Opt_D_dump_foreign

Opt_D_dump_inlinings

Opt_D_dump_rule_firings

Opt_D_dump_rule_rewrites

Opt_D_dump_occur_anal

Opt_D_dump_parsed

Opt_D_dump_rn

Opt_D_dump_core_pipeline

Opt_D_dump_simpl

Opt_D_dump_simpl_iterations

Opt_D_dump_simpl_phases

Opt_D_dump_spec

Opt_D_dump_prep

Opt_D_dump_stg

Opt_D_dump_stranal

Opt_D_dump_tc

Opt_D_dump_types

Opt_D_dump_rules

Opt_D_dump_cse

Opt_D_dump_worker_wrapper

Opt_D_dump_rn_trace

Opt_D_dump_rn_stats

Opt_D_dump_opt_cmm

Opt_D_dump_simpl_stats

Opt_D_dump_cs_trace

Opt_D_dump_tc_trace

Opt_D_dump_if_trace

Opt_D_dump_vt_trace

Opt_D_dump_splices

Opt_D_dump_BCOs

Opt_D_dump_vect

Opt_D_dump_ticked

Opt_D_dump_rtti

Opt_D_source_stats

Opt_D_verbose_core2core

Opt_D_verbose_stg2stg

Opt_D_dump_hi

Opt_D_dump_hi_diffs

Opt_D_dump_minimal_imports

Opt_D_dump_mod_cycles

Opt_D_dump_view_pattern_commoning

Opt_D_faststring_stats

Opt_DumpToFile
Append dump output to files instead of stdout.

Opt_D_no_debug_output

Opt_DoCoreLinting

Opt_DoStgLinting

Opt_DoCmmLinting

Opt_DoAsmLinting

Opt_NoLlvmMangler

Opt_WarnIsError

Opt_PrintExplicitForalls

Opt_Strictness

Opt_FullLaziness

Opt_FloatIn

Opt_Specialise

Opt_StaticArgumentTransformation

Opt_CSE

Opt_LiberateCase

Opt_SpecConstr

Opt_DoLambdaEtaExpansion

Opt_IgnoreAsserts

Opt_DoEtaReduction

Opt_CaseMerge

Opt_UnboxStrictFields

Opt_DictsCheap

Opt_EnableRewriteRules

Opt_Vectorise

Opt_RegsGraph

Opt_RegsIterative

Opt_PedanticBottoms

Opt_IgnoreInterfacePragmas

Opt_OmitInterfacePragmas

Opt_ExposeAllUnfoldings

Opt_AutoSccsOnIndividualCafs

Opt_ProfCountEntries

Opt_Pp

Opt_ForceRecomp

Opt_ExcessPrecision

Opt_EagerBlackHoling

Opt_ReadUserPackageConf

Opt_NoHsMain

Opt_SplitObjs

Opt_StgStats

Opt_HideAllPackages

Opt_PrintBindResult

Opt_Haddock

Opt_HaddockOptions

Opt_Hpc_No_Auto

Opt_BreakOnException

Opt_BreakOnError

Opt_PrintEvldWithShow

Opt_PrintBindContents

Opt_GenManifest

Opt_EmbedManifest

Opt_EmitExternalCore

Opt_SharedImplib

Opt_BuildingCabalPackage

Opt_SSE2

Opt_SSE4_2

Opt_GhciSandbox

Opt_GhciHistory

Opt_HelpfulErrors

Opt_RunCPS

Opt_RunCPSZ

Opt_AutoLinkPackages

Opt_ImplicitImportQualified

Opt_TryNewCodeGen

Opt_KeepHiDiffs

Opt_KeepHcFiles

Opt_KeepSFiles

Opt_KeepTmpFiles

Opt_KeepRawTokenStream

Opt_KeepLlvmFiles

Opt_DistrustAllPackages

Opt_PackageTrust

data Severity

Constructors

SevOutput

SevInfo

SevWarning

SevError

SevFatal

data HscTarget

The target code type of the compilation (if any).

Whenever you change the target, also make sure to set ghcLink to something sensible.

HscNothing can be used to avoid generating any output, however, note that:

This will not run the desugaring step, thus no warnings generated in this step will be output. In particular, this includes warnings related to pattern matching. You can run the desugarer manually using desugarModule.

If a program uses Template Haskell the typechecker may try to run code from an imported module. This will fail if no code has been generated for this module. You can use needsTemplateHaskell to detect whether this might be the case and choose to either switch to a different target or avoid typechecking such modules. (The latter may preferable for security reasons.)

Constructors

HscC
Generate C code.

HscAsm
Generate assembly using the native code generator.

HscLlvm
Generate assembly using the llvm code generator.

HscInterpreted
Generate bytecode. (Requires LinkInMemory)

HscNothing
Don't generate any code. See notes above.

dopt :: DynFlag -> DynFlags -> Bool

Test whether a DynFlag is set

data GhcMode

The GhcMode tells us whether we're doing multi-module compilation (controlled via the GHC API) or one-shot (single-module) compilation. This makes a difference primarily to the Finder: in one-shot mode we look for interface files for imported modules, but in multi-module mode we look for source files in order to check whether they need to be recompiled.

Constructors

CompManager
--make, GHCi, etc.

OneShot
ghc -c Foo.hs

MkDepend
ghc -M, see Finder for why we need this

data GhcLink

What to do in the link step, if there is one.

Constructors

NoLink
Don't link at all

LinkBinary
Link object code into a binary

LinkInMemory
Use the in-memory dynamic linker (works for both bytecode and object code).

LinkDynLib
Link objects into a dynamic lib (DLL on Windows, DSO on ELF platforms)

defaultObjectTarget :: HscTarget

The HscTarget value corresponding to the default way to create object files on the current platform.

parseDynamicFlags :: Monad m => DynFlags -> [Located String] -> m (DynFlags, [Located String], [Located String])

getSessionDynFlags :: GhcMonad m => m DynFlags

Grabs the DynFlags from the Session

setSessionDynFlags :: GhcMonad m => DynFlags -> m [PackageId]

Updates the DynFlags in a Session. This also reads the package database (unless it has already been read), and prepares the compilers knowledge about packages. It can be called again to load new packages: just add new package flags to (packageFlags dflags).

Returns a list of new packages that may need to be linked in using the dynamic linker (see linkPackages) as a result of new package flags. If you are not doing linking or doing static linking, you can ignore the list of packages returned.

parseStaticFlags :: [Located String] -> IO ([Located String], [Located String])

Parses GHC's static flags from a list of command line arguments.

These flags are static in the sense that they can be set only once and they are global, meaning that they affect every instance of GHC running; multiple GHC threads will use the same flags.

This function must be called before any session is started, i.e., before the first call to withGhc.

Static flags are more of a hack and are static for more or less historical reasons. In the long run, most static flags should eventually become dynamic flags.

XXX: can we add an auto-generated list of static flags here?

Targets

data Target

A compilation target.

A target may be supplied with the actual text of the module. If so, use this instead of the file contents (this is for use in an IDE where the file hasn't been saved by the user yet).

Constructors

Target

Fields

targetId :: TargetId

module or filename

targetAllowObjCode :: Bool

object code allowed?

targetContents :: Maybe (StringBuffer, ClockTime)

in-memory text buffer?

data TargetId

Constructors

TargetModule ModuleName
A module name: search for the file

TargetFile FilePath (Maybe Phase)
A filename: preprocess & parse it to find the module name. If specified, the Phase indicates how to compile this file (which phase to start from). Nothing indicates the starting phase should be determined from the suffix of the filename.

data Phase

setTargets :: GhcMonad m => [Target] -> m ()

Sets the targets for this session. Each target may be a module name or a filename. The targets correspond to the set of root modules for the program/library. Unloading the current program is achieved by setting the current set of targets to be empty, followed by load.

getTargets :: GhcMonad m => m [Target]

Returns the current set of targets

addTarget :: GhcMonad m => Target -> m ()

Add another target.

removeTarget :: GhcMonad m => TargetId -> m ()

Remove a target

guessTarget :: GhcMonad m => String -> Maybe Phase -> m Target

Attempts to guess what Target a string refers to. This function implements the --make/GHCi command-line syntax for filenames:

if the string looks like a Haskell source filename, then interpret it as such

if adding a .hs or .lhs suffix yields the name of an existing file, then use that

otherwise interpret the string as a module name

Loading/compiling the program

depanal

Arguments

:: GhcMonad m

=> [ModuleName]
excluded modules

-> Bool
allow duplicate roots

-> m ModuleGraph

Perform a dependency analysis starting from the current targets and update the session with the new module graph.

Dependency analysis entails parsing the import directives and may therefore require running certain preprocessors.

Note that each ModSummary in the module graph caches its DynFlags. These DynFlags are determined by the current session DynFlags and the OPTIONS and LANGUAGE pragmas of the parsed module. Thus if you want to changes to the DynFlags to take effect you need to call this function again.

load :: GhcMonad m => LoadHowMuch -> m SuccessFlag

Try to load the program. See LoadHowMuch for the different modes.

This function implements the core of GHC's --make mode. It preprocesses, compiles and loads the specified modules, avoiding re-compilation wherever possible. Depending on the target (see hscTarget) compilating and loading may result in files being created on disk.

Calls the reportModuleCompilationResult callback after each compiling each module, whether successful or not.

Throw a SourceError if errors are encountered before the actual compilation starts (e.g., during dependency analysis). All other errors are reported using the callback.

data LoadHowMuch

Describes which modules of the module graph need to be loaded.

Constructors

LoadAllTargets
Load all targets and its dependencies.

LoadUpTo ModuleName
Load only the given module and its dependencies.

LoadDependenciesOf ModuleName
Load only the dependencies of the given module, but not the module itself.

data InteractiveImport

Constructors

IIDecl (ImportDecl RdrName)
Bring the exports of a particular module (filtered by an import decl) into scope

IIModule Module
Bring into scope the entire top-level envt of of this module, including the things imported into it.

data SuccessFlag

Constructors

Succeeded

Failed

succeeded :: SuccessFlag -> Bool

failed :: SuccessFlag -> Bool

defaultWarnErrLogger :: WarnErrLogger

type WarnErrLogger = GhcMonad m => Maybe SourceError -> m ()

A function called to log warnings and errors.

workingDirectoryChanged :: GhcMonad m => m ()

Inform GHC that the working directory has changed. GHC will flush its cache of module locations, since it may no longer be valid.

Note: Before changing the working directory make sure all threads running in the same session have stopped. If you change the working directory, you should also unload the current program (set targets to empty, followed by load).

parseModule :: GhcMonad m => ModSummary -> m ParsedModule

Parse a module.

Throws a SourceError on parse error.

typecheckModule :: GhcMonad m => ParsedModule -> m TypecheckedModule

Typecheck and rename a parsed module.

Throws a SourceError if either fails.

desugarModule :: GhcMonad m => TypecheckedModule -> m DesugaredModule

Desugar a typechecked module.

loadModule :: (TypecheckedMod mod, GhcMonad m) => mod -> m mod

Load a module. Input doesn't need to be desugared.

A module must be loaded before dependent modules can be typechecked. This always includes generating a ModIface and, depending on the hscTarget, may also include code generation.

This function will always cause recompilation and will always overwrite previous compilation results (potentially files on disk).

data ParsedModule

The result of successful parsing.

Constructors

ParsedModule

Fields

pm_mod_summary :: ModSummary

pm_parsed_source :: ParsedSource

pm_extra_src_files :: [FilePath]

data TypecheckedModule

The result of successful typechecking. It also contains the parser result.

Constructors

TypecheckedModule

Fields

tm_parsed_module :: ParsedModule

tm_renamed_source :: Maybe RenamedSource

tm_typechecked_source :: TypecheckedSource

tm_checked_module_info :: ModuleInfo

tm_internals_ :: (TcGblEnv, ModDetails)

data DesugaredModule

The result of successful desugaring (i.e., translation to core). Also contains all the information of a typechecked module.

Constructors

DesugaredModule

Fields

dm_typechecked_module :: TypecheckedModule

dm_core_module :: ModGuts

type TypecheckedSource = LHsBinds Id

type ParsedSource = Located (HsModule RdrName)

type RenamedSource = (HsGroup Name, [LImportDecl Name], Maybe [LIE Name], Maybe LHsDocString)

class ParsedMod m => TypecheckedMod m where

Methods

renamedSource :: m -> Maybe RenamedSource

typecheckedSource :: m -> TypecheckedSource

moduleInfo :: m -> ModuleInfo

class ParsedMod m where

Methods

parsedSource :: m -> ParsedSource

coreModule :: DesugaredMod m => m -> ModGuts

Compiling to Core

data CoreModule

A CoreModule consists of just the fields of a ModGuts that are needed for the compileToCoreModule interface.

Constructors

CoreModule

Fields

cm_module :: !Module

Module name

cm_types :: !TypeEnv

Type environment for types declared in this module

cm_binds :: CoreProgram

Declarations

cm_safe :: SafeHaskellMode

Safe Haskell mode

compileToCoreModule :: GhcMonad m => FilePath -> m CoreModule

This is the way to get access to the Core bindings corresponding to a module. compileToCore parses, typechecks, and desugars the module, then returns the resulting Core module (consisting of the module name, type declarations, and function declarations) if successful.

compileToCoreSimplified :: GhcMonad m => FilePath -> m CoreModule

Like compileToCoreModule, but invokes the simplifier, so as to return simplified and tidied Core.

compileCoreToObj :: GhcMonad m => Bool -> CoreModule -> m ()

Takes a CoreModule and compiles the bindings therein to object code. The first argument is a bool flag indicating whether to run the simplifier. The resulting .o, .hi, and executable files, if any, are stored in the current directory, and named according to the module name. This has only so far been tested with a single self-contained module.

Inspecting the module structure of the program

type ModuleGraph = [ModSummary]

A ModuleGraph contains all the nodes from the home package (only). There will be a node for each source module, plus a node for each hi-boot module.

The graph is not necessarily stored in topologically-sorted order. Use topSortModuleGraph and flattenSCC to achieve this.

data ModSummary

A single node in a 'ModuleGraph. The nodes of the module graph are one of:

A regular Haskell source module

A hi-boot source module

An external-core source module

Constructors

ModSummary

Fields

ms_mod :: Module

Identity of the module

ms_hsc_src :: HscSource

The module source either plain Haskell, hs-boot or external core

ms_location :: ModLocation

Location of the various files belonging to the module

ms_hs_date :: ClockTime

Timestamp of source file

ms_obj_date :: Maybe ClockTime

Timestamp of object, if we have one

ms_srcimps :: [Located (ImportDecl RdrName)]

Source imports of the module

ms_textual_imps :: [Located (ImportDecl RdrName)]

Non-source imports of the module from the module *text*

ms_hspp_file :: FilePath

Filename of preprocessed source file

ms_hspp_opts :: DynFlags

Cached flags from OPTIONS, INCLUDE and LANGUAGE pragmas in the modules source code

ms_hspp_buf :: Maybe StringBuffer

The actual preprocessed source, if we have it

ms_mod_name :: ModSummary -> ModuleName

data ModLocation

Where a module lives on the file system: the actual locations of the .hs, .hi and .o files, if we have them

Constructors

ModLocation

Fields

ml_hs_file :: Maybe FilePath

ml_hi_file :: FilePath

ml_obj_file :: FilePath

getModSummary :: GhcMonad m => ModuleName -> m ModSummary

Return the ModSummary of a module with the given name.

The module must be part of the module graph (see hsc_mod_graph and ModuleGraph). If this is not the case, this function will throw a GhcApiError.

This function ignores boot modules and requires that there is only one non-boot module with the given name.

getModuleGraph :: GhcMonad m => m ModuleGraph

Get the module dependency graph.

isLoaded :: GhcMonad m => ModuleName -> m Bool

Return True == module is loaded.

topSortModuleGraph

Arguments

:: Bool
Drop hi-boot nodes? (see below)

-> [ModSummary]

-> Maybe ModuleName
Root module name. If Nothing, use the full graph.

-> [SCC ModSummary]

Calculate SCCs of the module graph, possibly dropping the hi-boot nodes The resulting list of strongly-connected-components is in topologically sorted order, starting with the module(s) at the bottom of the dependency graph (ie compile them first) and ending with the ones at the top.

Drop hi-boot nodes (first boolean arg)?

False: treat the hi-boot summaries as nodes of the graph, so the graph must be acyclic

True: eliminate the hi-boot nodes, and instead pretend the a source-import of Foo is an import of Foo The resulting graph has no hi-boot nodes, but can be cyclic

Inspecting modules

data ModuleInfo

Container for information about a Module.

getModuleInfo :: GhcMonad m => Module -> m (Maybe ModuleInfo)

Request information about a loaded Module

modInfoTyThings :: ModuleInfo -> [TyThing]

The list of top-level entities defined in a module

modInfoTopLevelScope :: ModuleInfo -> Maybe [Name]

modInfoExports :: ModuleInfo -> [Name]

modInfoInstances :: ModuleInfo -> [Instance]

Returns the instances defined by the specified module. Warning: currently unimplemented for package modules.

modInfoIsExportedName :: ModuleInfo -> Name -> Bool

modInfoLookupName :: GhcMonad m => ModuleInfo -> Name -> m (Maybe TyThing)

modInfoIface :: ModuleInfo -> Maybe ModIface

modInfoSafe :: ModuleInfo -> SafeHaskellMode

Retrieve module safe haskell mode

lookupGlobalName :: GhcMonad m => Name -> m (Maybe TyThing)

Looks up a global name: that is, any top-level name in any visible module. Unlike lookupName, lookupGlobalName does not use the interactive context, and therefore does not require a preceding setContext.

findGlobalAnns :: (GhcMonad m, Typeable a) => ([Word8] -> a) -> AnnTarget Name -> m [a]

mkPrintUnqualifiedForModule :: GhcMonad m => ModuleInfo -> m (Maybe PrintUnqualified)

data ModIface

A ModIface plus a ModDetails summarises everything we know about a compiled module. The ModIface is the stuff *before* linking, and can be written out to an interface file. The 'ModDetails is after linking and can be completely recovered from just the ModIface.

When we read an interface file, we also construct a ModIface from it, except that we explicitly make the mi_decls and a few other fields empty; as when reading we consolidate the declarations etc. into a number of indexed maps and environments in the ExternalPackageState.

Constructors

ModIface

Fields

mi_module :: !Module

Name of the module we are for

mi_iface_hash :: !Fingerprint

Hash of the whole interface

mi_mod_hash :: !Fingerprint

Hash of the ABI only

mi_flag_hash :: !Fingerprint

Hash of the important flags used when compiling this module

mi_orphan :: !WhetherHasOrphans

Whether this module has orphans

mi_finsts :: !WhetherHasFamInst

Whether this module has family instances

mi_boot :: !IsBootInterface

Read from an hi-boot file?

mi_deps :: Dependencies

The dependencies of the module. This is consulted for directly-imported modules, but not for anything else (hence lazy)

mi_usages :: [Usage]

Usages; kept sorted so that it's easy to decide whether to write a new iface file (changing usages doesn't affect the hash of this module) NOT STRICT! we read this field lazily from the interface file It is *only* consulted by the recompilation checker

mi_exports :: ![IfaceExport]

Exports Kept sorted by (mod,occ), to make version comparisons easier Records the modules that are the declaration points for things exported by this module, and the OccNames of those things

mi_exp_hash :: !Fingerprint

Hash of export list

mi_used_th :: !Bool

Module required TH splices when it was compiled. This disables recompilation avoidance (see #481).

mi_fixities :: [(OccName, Fixity)]

Fixities NOT STRICT! we read this field lazily from the interface file

mi_warns :: Warnings

Warnings NOT STRICT! we read this field lazily from the interface file

mi_anns :: [IfaceAnnotation]

Annotations NOT STRICT! we read this field lazily from the interface file

mi_decls :: [(Fingerprint, IfaceDecl)]

Type, class and variable declarations The hash of an Id changes if its fixity or deprecations change (as well as its type of course) Ditto data constructors, class operations, except that the hash of the parent class/tycon changes

mi_globals :: !(Maybe GlobalRdrEnv)

Binds all the things defined at the top level in the original source code for this module. which is NOT the same as mi_exports, nor mi_decls (which may contains declarations for things not actually defined by the user). Used for GHCi and for inspecting the contents of modules via the GHC API only.

(We need the source file to figure out the top-level environment, if we didn't compile this module from source then this field contains Nothing).

Strictly speaking this field should live in the HomeModInfo, but that leads to more plumbing.

mi_insts :: [IfaceInst]

Sorted class instance

mi_fam_insts :: [IfaceFamInst]

Sorted family instances

mi_rules :: [IfaceRule]

Sorted rules

mi_orphan_hash :: !Fingerprint

Hash for orphan rules, class and family instances, and vectorise pragmas combined

mi_vect_info :: !IfaceVectInfo

Vectorisation information

mi_warn_fn :: Name -> Maybe WarningTxt

Cached lookup for mi_warns

mi_fix_fn :: OccName -> Fixity

Cached lookup for mi_fixities

mi_hash_fn :: OccName -> Maybe (OccName, Fingerprint)

Cached lookup for mi_decls. The Nothing in mi_hash_fn means that the thing isn't in decls. It's useful to know that when seeing if we are up to date wrt. the old interface. The OccName is the parent of the name, if it has one.

mi_hpc :: !AnyHpcUsage

True if this program uses Hpc at any point in the program.

mi_trust :: !IfaceTrustInfo

Safe Haskell Trust information for this module.

mi_trust_pkg :: !Bool

Do we require the package this module resides in be trusted to trust this module? This is used for the situation where a module is Safe (so doesn't require the package be trusted itself) but imports some trustworthy modules from its own package (which does require its own package be trusted). See Note [RnNames . Trust Own Package]

data SafeHaskellMode

The various Safe Haskell modes

Constructors

Sf_None

Sf_Unsafe

Sf_Trustworthy

Sf_Safe

Sf_SafeInfered

Querying the environment

packageDbModules

Arguments

:: GhcMonad m

=> Bool
Only consider exposed packages.

-> m [Module]

Return all external modules available in the package database. Modules from the current session (i.e., from the HomePackageTable) are not included.

Printing

type PrintUnqualified = (QueryQualifyName, QueryQualifyModule)

alwaysQualify :: PrintUnqualified

Interactive evaluation

getBindings :: GhcMonad m => m [TyThing]

Return the bindings for the current interactive session.

getInsts :: GhcMonad m => m ([Instance], [FamInst])

Return the instances for the current interactive session.

getPrintUnqual :: GhcMonad m => m PrintUnqualified

findModule :: GhcMonad m => ModuleName -> Maybe FastString -> m Module

Takes a ModuleName and possibly a PackageId, and consults the filesystem and package database to find the corresponding Module, using the algorithm that is used for an import declaration.

lookupModule :: GhcMonad m => ModuleName -> Maybe FastString -> m Module

Like findModule, but differs slightly when the module refers to a source file, and the file has not been loaded via load. In this case, findModule will throw an error (module not loaded), but lookupModule will check to see whether the module can also be found in a package, and if so, that package Module will be returned. If not, the usual module-not-found error will be thrown.

setContext :: GhcMonad m => [InteractiveImport] -> m ()

Set the interactive evaluation context.

Setting the context doesn't throw away any bindings; the bindings we've built up in the InteractiveContext simply move to the new module. They always shadow anything in scope in the current context.

getContext :: GhcMonad m => m [InteractiveImport]

Get the interactive evaluation context, consisting of a pair of the set of modules from which we take the full top-level scope, and the set of modules from which we take just the exports respectively.

getNamesInScope :: GhcMonad m => m [Name]

Returns all names in scope in the current interactive context

getRdrNamesInScope :: GhcMonad m => m [RdrName]

getGRE :: GhcMonad m => m GlobalRdrEnv

get the GlobalRdrEnv for a session

moduleIsInterpreted :: GhcMonad m => Module -> m Bool

Returns True if the specified module is interpreted, and hence has its full top-level scope available.

getInfo :: GhcMonad m => Name -> m (Maybe (TyThing, Fixity, [Instance]))

Looks up an identifier in the current interactive context (for :info) Filter the instances by the ones whose tycons (or clases resp) are in scope (qualified or otherwise). Otherwise we list a whole lot too many! The exact choice of which ones to show, and which to hide, is a judgement call. (see Trac #1581)

exprType :: GhcMonad m => String -> m Type

Get the type of an expression

typeKind :: GhcMonad m => Bool -> String -> m (Type, Kind)

Get the kind of a type

parseName :: GhcMonad m => String -> m [Name]

Parses a string as an identifier, and returns the list of Names that the identifier can refer to in the current interactive context.

data RunResult

Constructors

RunOk [Name]
names bound by this evaluation

RunException SomeException
statement raised an exception

RunBreak ThreadId [Name] (Maybe BreakInfo)

runStmt :: GhcMonad m => String -> SingleStep -> m RunResult

Run a statement in the current interactive context. Statement may bind multple values.

runStmtWithLocation :: GhcMonad m => String -> Int -> String -> SingleStep -> m RunResult

Run a statement in the current interactive context. Passing debug information Statement may bind multple values.

runDecls :: GhcMonad m => String -> m [Name]

runDeclsWithLocation :: GhcMonad m => String -> Int -> String -> m [Name]

parseImportDecl :: GhcMonad m => String -> m (ImportDecl RdrName)

data SingleStep

Constructors

RunToCompletion

SingleStep

RunAndLogSteps

resume :: GhcMonad m => (SrcSpan -> Bool) -> SingleStep -> m RunResult

data Resume

data History

getHistorySpan :: GhcMonad m => History -> m SrcSpan

getHistoryModule :: History -> Module

getResumeContext :: GhcMonad m => m [Resume]

abandon :: GhcMonad m => m Bool

abandonAll :: GhcMonad m => m Bool

back :: GhcMonad m => m ([Name], Int, SrcSpan)

forward :: GhcMonad m => m ([Name], Int, SrcSpan)

showModule :: GhcMonad m => ModSummary -> m String

isModuleInterpreted :: GhcMonad m => ModSummary -> m Bool

compileExpr :: GhcMonad m => String -> m HValue

data HValue

dynCompileExpr :: GhcMonad m => String -> m Dynamic

obtainTermFromId :: GhcMonad m => Int -> Bool -> Id -> m Term

obtainTermFromVal :: GhcMonad m => Int -> Bool -> Type -> a -> m Term

reconstructType :: HscEnv -> Int -> Id -> IO (Maybe Type)

modInfoModBreaks :: ModuleInfo -> ModBreaks

data ModBreaks

All the information about the breakpoints for a given module

Constructors

ModBreaks

Fields

modBreaks_flags :: BreakArray

The array of flags, one per breakpoint, indicating which breakpoints are enabled.

modBreaks_locs :: !(Array BreakIndex SrcSpan)

An array giving the source span of each breakpoint.

modBreaks_vars :: !(Array BreakIndex [OccName])

An array giving the names of the free variables at each breakpoint.

modBreaks_decls :: !(Array BreakIndex [String])

An array giving the names of the declarations enclosing each breakpoint.

type BreakIndex = Int

Breakpoint index

data BreakInfo

data BreakArray

setBreakOn :: BreakArray -> Int -> IO Bool

setBreakOff :: BreakArray -> Int -> IO Bool

getBreak :: BreakArray -> Int -> IO (Maybe Word)

lookupName :: GhcMonad m => Name -> m (Maybe TyThing)

Returns the TyThing for a Name. The Name may refer to any entity known to GHC, including Names defined using runStmt.

Abstract syntax elements

Packages

data PackageId

Essentially just a string identifying a package, including the version: e.g. parsec-1.0

Modules

data Module

A Module is a pair of a PackageId and a ModuleName.

mkModule :: PackageId -> ModuleName -> Module

pprModule :: Module -> SDoc

moduleName :: Module -> ModuleName

modulePackageId :: Module -> PackageId

data ModuleName

A ModuleName is essentially a simple string, e.g. Data.List.

mkModuleName :: String -> ModuleName

moduleNameString :: ModuleName -> String

Names

data Name

A unique, unambigious name for something, containing information about where that thing originated.

isExternalName :: Name -> Bool

nameModule :: Name -> Module

pprParenSymName :: NamedThing a => a -> SDoc

print a NamedThing, adding parentheses if the name is an operator.

nameSrcSpan :: Name -> SrcSpan

class NamedThing a where

A class allowing convenient access to the Name of various datatypes

Methods

getOccName :: a -> OccName

getName :: a -> Name

data RdrName

Do not use the data constructors of RdrName directly: prefer the family of functions that creates them, such as mkRdrUnqual

Constructors

Unqual OccName
Used for ordinary, unqualified occurrences, e.g. x, y or Foo. Create such a RdrName with mkRdrUnqual

Qual ModuleName OccName
A qualified name written by the user in source code. The module isn't necessarily the module where the thing is defined; just the one from which it is imported. Examples are Bar.x, Bar.y or Bar.Foo. Create such a RdrName with mkRdrQual

Identifiers

type Id = Var

idType :: Id -> Kind

isImplicitId :: Id -> Bool

isImplicitId tells whether an Ids info is implied by other declarations, so we don't need to put its signature in an interface file, even if it's mentioned in some other interface unfolding.

isDeadBinder :: Id -> Bool

isExportedId :: Var -> Bool

isExportedIdVar means "don't throw this away"

isLocalId :: Var -> Bool

isGlobalId :: Var -> Bool

isRecordSelector :: Id -> Bool

isPrimOpId :: Id -> Bool

isFCallId :: Id -> Bool

isClassOpId_maybe :: Id -> Maybe Class

isDataConWorkId :: Id -> Bool

idDataCon :: Id -> DataCon

Get from either the worker or the wrapper Id to the DataCon. Currently used only in the desugarer.

INVARIANT: idDataCon (dataConWrapId d) = d: remember, dataConWrapId can return either the wrapper or the worker

isBottomingId :: Id -> Bool

Returns true if an application to n args would diverge

isDictonaryId :: Id -> Bool

recordSelectorFieldLabel :: Id -> (TyCon, FieldLabel)

If the Id is that for a record selector, extract the sel_tycon and label. Panic otherwise

Type constructors

data TyCon

TyCons represent type constructors. Type constructors are introduced by things such as:

1) Data declarations: data Foo = ... creates the Foo type constructor of kind *

2) Type synonyms: type Foo = ... creates the Foo type constructor

3) Newtypes: newtype Foo a = MkFoo ... creates the Foo type constructor of kind * -> *

4) Class declarations: class Foo where creates the Foo type constructor of kind *

This data type also encodes a number of primitive, built in type constructors such as those for function and tuple types.

tyConTyVars :: TyCon -> [TyVar]

The kind and type variables used in the type constructor. Invariant: length tyvars = arity Precisely, this list scopes over:

The algTcStupidTheta 2. The cached types in 'algTyConRhs.NewTyCon' 3. The family instance types if present

Note that it does not scope over the data constructors.

tyConDataCons :: TyCon -> [DataCon]

As tyConDataCons_maybe, but returns the empty list of constructors if no constructors could be found

tyConArity :: TyCon -> Arity

n if ty_con :: * -> ... -> * n times

isClassTyCon :: TyCon -> Bool

Is this TyCon that for a class instance?

isSynTyCon :: TyCon -> Bool

Is this a TyCon representing a type synonym (type)?

isNewTyCon :: TyCon -> Bool

Is this TyCon that for a newtype

isPrimTyCon :: TyCon -> Bool

Does this TyCon represent something that cannot be defined in Haskell?

isFunTyCon :: TyCon -> Bool

isFamilyTyCon :: TyCon -> Bool

Is this a TyCon, synonym or otherwise, that may have further instances appear?

tyConClass_maybe :: TyCon -> Maybe Class

If this TyCon is that for a class instance, return the class it is for. Otherwise returns Nothing

synTyConDefn :: TyCon -> ([TyVar], Type)

Extract the TyVars bound by a type synonym and the corresponding (unsubstituted) right hand side. If the given TyCon is not a type synonym, panics

synTyConType :: TyCon -> Type

Find the expansion of the type synonym represented by the given TyCon. The free variables of this type will typically include those TyVars bound by the TyCon. Panics if the TyCon is not that of a type synonym

synTyConResKind :: TyCon -> Kind

Find the result Kind of a type synonym, after applying it to its arity number of type variables Actually this function works fine on data types too, but they'd always return *, so we never need to ask

Type variables

type TyVar = Var

alphaTyVars :: [TyVar]

Data constructors

data DataCon

A data constructor

dataConSig :: DataCon -> ([TyVar], ThetaType, [Type], Type)

The "signature" of the DataCon returns, in order:

1) The result of dataConAllTyVars,

2) All the ThetaTypes relating to the DataCon (coercion, dictionary, implicit parameter - whatever)

3) The type arguments to the constructor

4) The original result type of the DataCon

dataConType :: DataCon -> Type

dataConTyCon :: DataCon -> TyCon

The type constructor that we are building via this data constructor

dataConFieldLabels :: DataCon -> [FieldLabel]

The labels for the fields of this particular DataCon

dataConIsInfix :: DataCon -> Bool

Should the DataCon be presented infix?

isVanillaDataCon :: DataCon -> Bool

Vanilla DataCons are those that are nice boring Haskell 98 constructors

dataConUserType :: DataCon -> Type

The user-declared type of the data constructor in the nice-to-read form:

 T :: forall a b. a -> b -> T [a]

rather than:

 T :: forall a c. forall b. (c~[a]) => a -> b -> T c

NB: If the constructor is part of a data instance, the result type mentions the family tycon, not the internal one.

dataConStrictMarks :: DataCon -> [HsBang]

The strictness markings decided on by the compiler. Does not include those for existential dictionaries. The list is in one-to-one correspondence with the arity of the DataCon

data StrictnessMark

Constructors

MarkedStrict

NotMarkedStrict

isMarkedStrict :: StrictnessMark -> Bool

Classes

data Class

classMethods :: Class -> [Id]

classSCTheta :: Class -> [PredType]

classTvsFds :: Class -> ([TyVar], [FunDep TyVar])

classATs :: Class -> [TyCon]

pprFundeps :: Outputable a => [FunDep a] -> SDoc

Instances

data Instance

instanceDFunId :: Instance -> DFunId

pprInstance :: Instance -> SDoc

pprInstanceHdr :: Instance -> SDoc

pprFamInst :: FamInst -> SDoc

pprFamInstHdr :: FamInst -> SDoc

Types and Kinds

data Type

The key representation of types within the compiler

splitForAllTys :: Type -> ([TyVar], Type)

Attempts to take a forall type apart, returning all the immediate such bound type variables and the remainder of the type. Always suceeds, even if that means returning an empty list of TyVars

funResultTy :: Type -> Type

Extract the function result type and panic if that is not possible

pprParendType :: Type -> SDoc

pprTypeApp :: NamedThing a => a -> [Type] -> SDoc

type Kind = Type

The key type representing kinds in the compiler. Invariant: a kind is always in one of these forms:

 FunTy k1 k2
 TyConApp PrimTyCon [...]
 TyVar kv   -- (during inference only)
 ForAll ... -- (for top-level coercions)

type PredType = Type

A type of the form p of kind Constraint represents a value whose type is the Haskell predicate p, where a predicate is what occurs before the => in a Haskell type.

We use PredType as documentation to mark those types that we guarantee to have this kind.

It can be expanded into its representation, but:

The type checker must treat it as opaque

The rest of the compiler treats it as transparent

Consider these examples:

 f :: (Eq a) => a -> Int
 g :: (?x :: Int -> Int) => a -> Int
 h :: (r\l) => {r} => {l::Int | r}

Here the Eq a and ?x :: Int -> Int and rl are all called "predicates"

type ThetaType = [PredType]

A collection of PredTypes

pprForAll :: [TyVar] -> SDoc

pprThetaArrowTy :: ThetaType -> SDoc

Entities

data TyThing

A typecheckable-thing, essentially anything that has a name

Constructors

AnId Id

ADataCon DataCon

ATyCon TyCon

ACoAxiom CoAxiom

Syntax

module HsSyn

Fixities

data FixityDirection

Constructors

InfixL

InfixR

InfixN

defaultFixity :: Fixity

maxPrecedence :: Int

negateFixity :: Fixity

compareFixity :: Fixity -> Fixity -> (Bool, Bool)

Source locations

data SrcLoc

Constructors

RealSrcLoc !RealSrcLoc

UnhelpfulLoc FastString

data RealSrcLoc

Represents a single point within a file

mkSrcLoc :: FastString -> Int -> Int -> SrcLoc

noSrcLoc :: SrcLoc

Built-in bad SrcLoc values for particular locations

srcLocFile :: RealSrcLoc -> FastString

Gives the filename of the RealSrcLoc

srcLocLine :: RealSrcLoc -> Int

Raises an error when used on a bad SrcLoc

srcLocCol :: RealSrcLoc -> Int

Raises an error when used on a bad SrcLoc

data SrcSpan

Constructors

RealSrcSpan !RealSrcSpan

UnhelpfulSpan !FastString

data RealSrcSpan

A SrcSpan delimits a portion of a text file. It could be represented by a pair of (line,column) coordinates, but in fact we optimise slightly by using more compact representations for single-line and zero-length spans, both of which are quite common.

The end position is defined to be the column after the end of the span. That is, a span of (1,1)-(1,2) is one character long, and a span of (1,1)-(1,1) is zero characters long.

mkSrcSpan :: SrcLoc -> SrcLoc -> SrcSpan

Create a SrcSpan between two points in a file

srcLocSpan :: SrcLoc -> SrcSpan

Create a SrcSpan corresponding to a single point

isGoodSrcSpan :: SrcSpan -> Bool

Test if a SrcSpan is good, i.e. has precise location information

noSrcSpan :: SrcSpan

Built-in bad SrcSpans for common sources of location uncertainty

srcSpanStart :: SrcSpan -> SrcLoc

Returns the location at the start of the SrcSpan or a bad SrcSpan if that is unavailable

srcSpanEnd :: SrcSpan -> SrcLoc

Returns the location at the end of the SrcSpan or a bad SrcSpan if that is unavailable

srcSpanFile :: RealSrcSpan -> FastString

srcSpanStartLine :: RealSrcSpan -> Int

srcSpanEndLine :: RealSrcSpan -> Int

srcSpanStartCol :: RealSrcSpan -> Int

srcSpanEndCol :: RealSrcSpan -> Int

Located

data GenLocated l e

We attach SrcSpans to lots of things, so let's have a datatype for it.

Constructors

L l e

type Located e = GenLocated SrcSpan e

Constructing Located

noLoc :: e -> Located e

mkGeneralLocated :: String -> e -> Located e

Deconstructing Located

getLoc :: GenLocated l e -> l

unLoc :: GenLocated l e -> e

Combining and comparing Located values

eqLocated :: Eq a => Located a -> Located a -> Bool

Tests whether the two located things are equal

cmpLocated :: Ord a => Located a -> Located a -> Ordering

Tests the ordering of the two located things

combineLocs :: Located a -> Located b -> SrcSpan

addCLoc :: Located a -> Located b -> c -> Located c

Combine locations from two Located things and add them to a third thing

leftmost_smallest :: SrcSpan -> SrcSpan -> Ordering

Alternative strategies for ordering SrcSpans

leftmost_largest :: SrcSpan -> SrcSpan -> Ordering

Alternative strategies for ordering SrcSpans

rightmost :: SrcSpan -> SrcSpan -> Ordering

Alternative strategies for ordering SrcSpans

spans :: SrcSpan -> (Int, Int) -> Bool

Determines whether a span encloses a given line and column index

isSubspanOf

Arguments

:: SrcSpan
The span that may be enclosed by the other

-> SrcSpan
The span it may be enclosed by

-> Bool

Determines whether a span is enclosed by another one

Exceptions

data GhcException

GHC's own exception type error messages all take the form:

	location: error

If the location is on the command line, or in GHC itself, then location=ghc. All of the error types below correspond to a location of ghc, except for ProgramError (where the string is assumed to contain a location already, so we don't print one).

Constructors

PhaseFailed String ExitCode

Signal Int
Some other fatal signal (SIGHUP,SIGTERM)

UsageError String
Prints the short usage msg after the error

CmdLineError String
A problem with the command line arguments, but don't print usage.

Panic String
The impossible happened.

Sorry String
The user tickled something that's known not to work yet, but we're not counting it as a bug.

InstallationError String
An installation problem.

ProgramError String
An error in the user's code, probably.

showGhcException :: GhcException -> String -> String

Append a description of the given exception to this string.

Token stream manipulations

data Token

getTokenStream :: GhcMonad m => Module -> m [Located Token]

Return module source as token stream, including comments.

The module must be in the module graph and its source must be available. Throws a SourceError on parse error.

getRichTokenStream :: GhcMonad m => Module -> m [(Located Token, String)]

Give even more information on the source than getTokenStream This function allows reconstructing the source completely with showRichTokenStream.

showRichTokenStream :: [(Located Token, String)] -> String

Take a rich token stream such as produced from getRichTokenStream and return source code almost identical to the original code (except for insignificant whitespace.)

addSourceToTokens :: RealSrcLoc -> StringBuffer -> [Located Token] -> [(Located Token, String)]

Given a source location and a StringBuffer corresponding to this location, return a rich token stream with the source associated to the tokens.

Pure interface to the parser

parser

Arguments

:: String
Haskell module source text (full Unicode is supported)

-> DynFlags
the flags

-> FilePath
the filename (for source locations)

-> Either ErrorMessages (WarningMessages, Located (HsModule RdrName))

A pure interface to the module parser.

Miscellaneous

cyclicModuleErr :: [ModSummary] -> SDoc