GHC.Builtin.Names

Documentation

data Unique

Unique identifier.

The type of unique identifiers that are used in many places in GHC for fast ordering and equality tests. You should generate these with the functions from the UniqSupply module

These are sometimes also referred to as "keys" in comments in GHC.

class Uniquable a where

Class of things that we can obtain a Unique from

Methods

getUnique :: a -> Unique

hasKey :: Uniquable a => a -> Unique -> Bool

aRROW :: Module

absentErrorIdKey :: Unique

absentSumFieldErrorIdKey :: Unique

addrPrimTyConKey :: Unique

allNameStrings :: [String]

all_RDR :: RdrName

alt_RDR :: RdrName

alternativeClassKey :: Unique

alternativeClassName :: Name

and_RDR :: RdrName

anyBoxConKey :: Unique

anyTyConKey :: Unique

apAClassOpKey :: Unique

apAName :: Name

ap_RDR :: RdrName

appAIdKey :: Unique

appAName :: Name

appendIdKey :: Unique

appendName :: Name

append_RDR :: RdrName

applicativeClassKey :: Unique

applicativeClassName :: Name

arrAIdKey :: Unique

arrAName :: Name

arrayArrayPrimTyConKey :: Unique

arrayPrimTyConKey :: Unique

assertErrorIdKey :: Unique

assertErrorName :: Name

assertIdKey :: Unique

assertName :: Name

augmentIdKey :: Unique

augmentName :: Name

basicKnownKeyNames :: [Name]

bcoPrimTyConKey :: Unique

bignatFromWordListIdKey :: Unique

bignatFromWordListName :: Name

bindIOIdKey :: Unique

bindIOName :: Name

bindIO_RDR :: RdrName

bindMClassOpKey :: Unique

bindMName :: Name

bindM_RDR :: RdrName

bitIntegerIdKey :: Unique

bnbVarQual :: String -> Unique -> Name

bniVarQual :: String -> Unique -> Name

bnnVarQual :: String -> Unique -> Name

boolTyConKey :: Unique

boundedClassKey :: Unique

boundedClassName :: Name

boxedRepDataConKey :: Unique

boxityConKey :: Unique

breakpointCondIdKey :: Unique

breakpointCondName :: Name

breakpointIdKey :: Unique

breakpointName :: Name

buildIdKey :: Unique

buildName :: Name

build_RDR :: RdrName

byteArrayPrimTyConKey :: Unique

c1TyConKey :: Unique

c1TyConName :: Name

cONTROL_APPLICATIVE :: Module

cONTROL_EXCEPTION_BASE :: Module

cTyConKey :: Unique

cTyConName :: Name

callStackTyConKey :: Unique

callStackTyConName :: Name

charDataConKey :: Unique

charPrimTyConKey :: Unique

charTyConKey :: Unique

choiceAIdKey :: Unique

choiceAName :: Name

choose_RDR :: RdrName

clsQual :: Module -> FastString -> Unique -> Name

clsQual_RDR :: Module -> FastString -> RdrName

coerceKey :: Unique

coercibleDataConKey :: Unique

coercibleSCSelIdKey :: Unique

coercibleTyConKey :: Unique

coercionTokenIdKey :: Unique

comp1DataCon_RDR :: RdrName

compTyConKey :: Unique

compTyConName :: Name

compactPrimTyConKey :: Unique

compare_RDR :: RdrName

composeAIdKey :: Unique

composeAName :: Name

compose_RDR :: RdrName

conFixity_RDR :: RdrName

conIsRecord_RDR :: RdrName

conName_RDR :: RdrName

concatIdKey :: Unique

concatName :: Name

consDataConKey :: Unique

considerAccessibleIdKey :: Unique

considerAccessibleName :: Name

constraintKindTyConKey :: Unique

constructorClassKey :: Unique

constructorClassName :: Name

crossDataConKey :: Unique

cstringLengthIdKey :: Unique

cstringLengthName :: Name

d1TyConKey :: Unique

d1TyConName :: Name

dATA_ARRAY_PARALLEL_NAME :: ModuleName

dATA_ARRAY_PARALLEL_PRIM_NAME :: ModuleName

dATA_COERCE :: Module

dATA_EITHER :: Module

dATA_FOLDABLE :: Module

dATA_LIST :: Module

dATA_STRING :: Module

dATA_TRAVERSABLE :: Module

dATA_TUPLE :: Module

dATA_TYPE_EQUALITY :: Module

dEBUG_TRACE :: Module

dTyConKey :: Unique

dTyConName :: Name

dYNAMIC :: Module

dataClassKey :: Unique

dataClassName :: Name

dataQual_RDR :: Module -> FastString -> RdrName

dataToTag_RDR :: RdrName

datatypeClassKey :: Unique

datatypeClassName :: Name

datatypeName_RDR :: RdrName

dcQual :: Module -> FastString -> Unique -> Name

decidedLazyDataConKey :: Unique

decidedLazyDataConName :: Name

decidedStrictDataConKey :: Unique

decidedStrictDataConName :: Name

decidedUnpackDataConKey :: Unique

decidedUnpackDataConName :: Name

derivableClassKeys :: [Unique]

divIntIdKey :: Unique

divIntName :: Name

dollarIdKey :: Unique

dollarName :: Name

doubleDataConKey :: Unique

doublePrimTyConKey :: Unique

doubleTyConKey :: Unique

doubleX2PrimTyConKey :: Unique

doubleX4PrimTyConKey :: Unique

doubleX8PrimTyConKey :: Unique

eitherTyConKey :: Unique

eitherTyConName :: Name

emptyCallStackKey :: Unique

emptyCallStackName :: Name

enumClassKey :: Unique

enumClassName :: Name

enumClass_RDR :: RdrName

enumFromClassOpKey :: Unique

enumFromName :: Name

enumFromThenClassOpKey :: Unique

enumFromThenName :: Name

enumFromThenToClassOpKey :: Unique

enumFromThenToName :: Name

enumFromThenTo_RDR :: RdrName

enumFromThen_RDR :: RdrName

enumFromToClassOpKey :: Unique

enumFromToName :: Name

enumFromTo_RDR :: RdrName

enumFrom_RDR :: RdrName

eqClassKey :: Unique

eqClassName :: Name

eqClassOpKey :: Unique

eqClass_RDR :: RdrName

eqDataConKey :: Unique

eqName :: Name

eqPhantPrimTyConKey :: Unique

eqPrimTyConKey :: Unique

eqReprPrimTyConKey :: Unique

eqSCSelIdKey :: Unique

eqStringIdKey :: Unique

eqStringName :: Name

eqString_RDR :: RdrName

eqTag_RDR :: RdrName

eqTyConKey :: Unique

eq_RDR :: RdrName

errorIdKey :: Unique

errorMessageTypeErrorFamKey :: Unique

errorMessageTypeErrorFamName :: Name

error_RDR :: RdrName

expectP_RDR :: RdrName

failIOIdKey :: Unique

failIOName :: Name

failMClassOpKey :: Unique

failMName :: Name

failM_RDR :: RdrName

falseDataConKey :: Unique

filterIdKey :: Unique

filterName :: Name

fingerprintDataConKey :: Unique

fingerprintDataConName :: Name

firstAIdKey :: Unique

firstAName :: Name

floatDataConKey :: Unique

floatPrimTyConKey :: Unique

floatTyConKey :: Unique

floatX16PrimTyConKey :: Unique

floatX4PrimTyConKey :: Unique

floatX8PrimTyConKey :: Unique

floatingClassKey :: Unique

floatingClassName :: Name

fmapClassOpKey :: Unique

fmapName :: Name

fmap_RDR :: RdrName

foldMap_RDR :: RdrName

foldableClassKey :: Unique

foldableClassName :: Name

foldable_foldr_RDR :: RdrName

foldrIdKey :: Unique

foldrName :: Name

foldr_RDR :: RdrName

foreignObjPrimTyConKey :: Unique

fractionalClassKey :: Unique

fractionalClassKeys :: [Unique]

fractionalClassName :: Name

from1_RDR :: RdrName

fromEnum_RDR :: RdrName

fromIntegerClassOpKey :: Unique

fromIntegerName :: Name

fromInteger_RDR :: RdrName

fromIntegralIdKey :: Unique

fromIntegralName :: Name

fromIntegral_RDR :: RdrName

fromLabelClassOpKey :: Unique

fromLabelClassOpName :: Name

fromListClassOpKey :: Unique

fromListNClassOpKey :: Unique

fromListNName :: Name

fromListN_RDR :: RdrName

fromListName :: Name

fromList_RDR :: RdrName

fromRationalClassOpKey :: Unique

fromRationalName :: Name

fromRational_RDR :: RdrName

fromStaticPtrClassOpKey :: Unique

fromStaticPtrName :: Name

fromStringClassOpKey :: Unique

fromStringName :: Name

fromString_RDR :: RdrName

from_RDR :: RdrName

frontendPluginTyConKey :: Unique

frontendPluginTyConName :: Name

fstIdKey :: Unique

fstName :: Name

funPtrTyConKey :: Unique

funPtrTyConName :: Name

funTyConKey :: Unique

functorClassKey :: Unique

functorClassName :: Name

gENERICS :: Module

gHC_BASE :: Module

gHC_CLASSES :: Module

gHC_CONC :: Module

gHC_CSTRING :: Module

gHC_DESUGAR :: Module

gHC_ENUM :: Module

gHC_ERR :: Module

gHC_EXTS :: Module

gHC_FINGERPRINT_TYPE :: Module

gHC_FLOAT :: Module

gHC_GENERICS :: Module

gHC_GHCI :: Module

gHC_GHCI_HELPERS :: Module

gHC_INT :: Module

gHC_IO :: Module

gHC_IO_Exception :: Module

gHC_IX :: Module

gHC_LIST :: Module

gHC_MAGIC :: Module

gHC_MAYBE :: Module

gHC_NUM :: Module

gHC_NUM_BIGNAT :: Module

gHC_NUM_INTEGER :: Module

gHC_NUM_NATURAL :: Module

gHC_OVER_LABELS :: Module

gHC_PRIM :: Module

gHC_PRIMOPWRAPPERS :: Module

gHC_PRIM_EXCEPTION :: Module

gHC_PRIM_PANIC :: Module

gHC_PTR :: Module

gHC_READ :: Module

gHC_REAL :: Module

gHC_RECORDS :: Module

gHC_SHOW :: Module

gHC_SRCLOC :: Module

gHC_ST :: Module

gHC_STABLE :: Module

gHC_STACK :: Module

gHC_STACK_TYPES :: Module

gHC_STATICPTR :: Module

gHC_STATICPTR_INTERNAL :: Module

gHC_TOP_HANDLER :: Module

gHC_TUPLE :: Module

gHC_TYPELITS :: Module

gHC_TYPELITS_INTERNAL :: Module

gHC_TYPENATS :: Module

gHC_TYPENATS_INTERNAL :: Module

gHC_TYPES :: Module

gHC_WORD :: Module

geClassOpKey :: Unique

geName :: Name

ge_RDR :: RdrName

gen1ClassKey :: Unique

gen1ClassName :: Name

genClassKey :: Unique

genClassName :: Name

genUnitDataConKey :: Unique

genericClassNames :: [Name]

genericTyConNames :: [Name]

getFieldClassOpKey :: Unique

getFieldName :: Name

getTag_RDR :: RdrName

ghciIoClassKey :: Unique

ghciIoClassName :: Name

ghciStepIoMClassOpKey :: Unique

ghciStepIoMName :: Name

groupWithIdKey :: Unique

groupWithName :: Name

gtTag_RDR :: RdrName

gt_RDR :: RdrName

guardMIdKey :: Unique

guardMName :: Name

hasFieldClassName :: Name

hasFieldClassNameKey :: Unique

heqDataConKey :: Unique

heqSCSelIdKey :: Unique

heqTyConKey :: Unique

ident_RDR :: RdrName

inRange_RDR :: RdrName

index_RDR :: RdrName

infixDataCon_RDR :: RdrName

infixIDataConKey :: Unique

infixIDataConName :: Name

inlDataConKey :: Unique

inlineIdKey :: Unique

inlineIdName :: Name

inrDataConKey :: Unique

int16PrimTyConKey :: Unique

int16TyConKey :: Unique

int16TyConName :: Name

int16X16PrimTyConKey :: Unique

int16X32PrimTyConKey :: Unique

int16X8PrimTyConKey :: Unique

int32PrimTyConKey :: Unique

int32TyConKey :: Unique

int32TyConName :: Name

int32X16PrimTyConKey :: Unique

int32X4PrimTyConKey :: Unique

int32X8PrimTyConKey :: Unique

int64PrimTyConKey :: Unique

int64TyConKey :: Unique

int64TyConName :: Name

int64X2PrimTyConKey :: Unique

int64X4PrimTyConKey :: Unique

int64X8PrimTyConKey :: Unique

int8PrimTyConKey :: Unique

int8TyConKey :: Unique

int8TyConName :: Name

int8X16PrimTyConKey :: Unique

int8X32PrimTyConKey :: Unique

int8X64PrimTyConKey :: Unique

intDataConKey :: Unique

intPrimTyConKey :: Unique

intTyConKey :: Unique

integerAbsIdKey :: Unique

integerAbsName :: Name

integerAddIdKey :: Unique

integerAddName :: Name

integerAdd_RDR :: RdrName

integerAndIdKey :: Unique

integerAndName :: Name

integerBitIdKey :: Unique

integerBitName :: Name

integerCompareIdKey :: Unique

integerCompareName :: Name

integerComplementIdKey :: Unique

integerComplementName :: Name

integerDivIdKey :: Unique

integerDivModIdKey :: Unique

integerDivModName :: Name

integerDivName :: Name

integerEncodeDoubleIdKey :: Unique

integerEncodeDoubleName :: Name

integerEncodeFloatIdKey :: Unique

integerEncodeFloatName :: Name

integerEqIdKey :: Unique

integerEqName :: Name

integerFromInt64IdKey :: Unique

integerFromInt64Name :: Name

integerFromNaturalIdKey :: Unique

integerFromNaturalName :: Name

integerFromWord64IdKey :: Unique

integerFromWord64Name :: Name

integerFromWordIdKey :: Unique

integerFromWordName :: Name

integerGcdIdKey :: Unique

integerGcdName :: Name

integerGeIdKey :: Unique

integerGeName :: Name

integerGtIdKey :: Unique

integerGtName :: Name

integerINDataConKey :: Unique

integerIPDataConKey :: Unique

integerISDataConKey :: Unique

integerLcmIdKey :: Unique

integerLcmName :: Name

integerLeIdKey :: Unique

integerLeName :: Name

integerLtIdKey :: Unique

integerLtName :: Name

integerModIdKey :: Unique

integerModName :: Name

integerMulIdKey :: Unique

integerMulName :: Name

integerMul_RDR :: RdrName

integerNeIdKey :: Unique

integerNeName :: Name

integerNegateIdKey :: Unique

integerNegateName :: Name

integerOrIdKey :: Unique

integerOrName :: Name

integerPopCountIdKey :: Unique

integerPopCountName :: Name

integerQuotIdKey :: Unique

integerQuotName :: Name

integerQuotRemIdKey :: Unique

integerQuotRemName :: Name

integerRemIdKey :: Unique

integerRemName :: Name

integerShiftLIdKey :: Unique

integerShiftLName :: Name

integerShiftRIdKey :: Unique

integerShiftRName :: Name

integerSignumIdKey :: Unique

integerSignumName :: Name

integerSubIdKey :: Unique

integerSubName :: Name

integerTestBitIdKey :: Unique

integerTestBitName :: Name

integerToDoubleIdKey :: Unique

integerToDoubleName :: Name

integerToFloatIdKey :: Unique

integerToFloatName :: Name

integerToInt64IdKey :: Unique

integerToInt64Name :: Name

integerToIntIdKey :: Unique

integerToIntName :: Name

integerToNaturalClampIdKey :: Unique

integerToNaturalClampName :: Name

integerToNaturalIdKey :: Unique

integerToNaturalName :: Name

integerToNaturalThrowIdKey :: Unique

integerToNaturalThrowName :: Name

integerToWord64IdKey :: Unique

integerToWord64Name :: Name

integerToWordIdKey :: Unique

integerToWordName :: Name

integerTyConKey :: Unique

integerXorIdKey :: Unique

integerXorName :: Name

integralClassKey :: Unique

integralClassName :: Name

interactiveClassKeys :: [Unique]

interactiveClassNames :: [Name]

ioDataConKey :: Unique

ioDataConName :: Name

ioDataCon_RDR :: RdrName

ioPortPrimTyConKey :: Unique

ioTyConKey :: Unique

ioTyConName :: Name

ipClassKey :: Unique

ipClassName :: Name

isListClassKey :: Unique

isListClassName :: Name

isNewtypeName_RDR :: RdrName

isStringClassKey :: Unique

isStringClassName :: Name

isUnboundName :: Name -> Bool

itName :: Unique -> SrcSpan -> Name

ixClassKey :: Unique

ixClassName :: Name

joinMIdKey :: Unique

joinMName :: Name

justDataConKey :: Unique

k1DataCon_RDR :: RdrName

k1TyConKey :: Unique

k1TyConName :: Name

kindConKey :: Unique

kindRepAppDataConKey :: Unique

kindRepAppDataConName :: Name

kindRepFunDataConKey :: Unique

kindRepFunDataConName :: Name

kindRepTYPEDataConKey :: Unique

kindRepTYPEDataConName :: Name

kindRepTyConAppDataConKey :: Unique

kindRepTyConAppDataConName :: Name

kindRepTyConKey :: Unique

kindRepTyConName :: Name

kindRepTypeLitDDataConKey :: Unique

kindRepTypeLitDDataConName :: Name

kindRepTypeLitSDataConKey :: Unique

kindRepTypeLitSDataConName :: Name

kindRepVarDataConKey :: Unique

kindRepVarDataConName :: Name

knownCharClassName :: Name

knownCharClassNameKey :: Unique

knownNatClassName :: Name

knownNatClassNameKey :: Unique

knownSymbolClassName :: Name

knownSymbolClassNameKey :: Unique

l1DataCon_RDR :: RdrName

lEX :: Module

lazyIdKey :: Unique

le_RDR :: RdrName

leftAssocDataCon_RDR :: RdrName

leftAssociativeDataConKey :: Unique

leftAssociativeDataConName :: Name

leftDataConKey :: Unique

leftDataConName :: Name

leftSectionKey :: Unique

left_RDR :: RdrName

levityTyConKey :: Unique

lexP_RDR :: RdrName

liftA2_RDR :: RdrName

liftMIdKey :: Unique

liftMName :: Name

liftedDataConKey :: Unique

liftedRepTyConKey :: Unique

liftedTypeKindTyConKey :: Unique

listTyConKey :: Unique

loopAIdKey :: Unique

loopAName :: Name

ltTag_RDR :: RdrName

lt_RDR :: RdrName

m1DataCon_RDR :: RdrName

m1TyConKey :: Unique

m1TyConName :: Name

mAIN_NAME :: ModuleName

mONAD :: Module

mONAD_FAIL :: Module

mONAD_FIX :: Module

mONAD_ZIP :: Module

mVarPrimTyConKey :: Unique

magicDictKey :: Unique

main_RDR_Unqual :: RdrName

makeStaticKey :: Unique

makeStaticName :: Name

manyDataConKey :: Unique

mapIdKey :: Unique

mapName :: Name

map_RDR :: RdrName

mappendClassOpKey :: Unique

mappendName :: Name

mappend_RDR :: RdrName

maxBound_RDR :: RdrName

maybeTyConKey :: Unique

mconcatClassOpKey :: Unique

mconcatName :: Name

memptyClassOpKey :: Unique

memptyName :: Name

mempty_RDR :: RdrName

metaConsDataConKey :: Unique

metaConsDataConName :: Name

metaDataDataConKey :: Unique

metaDataDataConName :: Name

metaSelDataConKey :: Unique

metaSelDataConName :: Name

mfixIdKey :: Unique

mfixName :: Name

minBound_RDR :: RdrName

minusClassOpKey :: Unique

minusName :: Name

minus_RDR :: RdrName

mkBaseModule :: FastString -> Module

mkBaseModule_ :: ModuleName -> Module

mkBignumModule :: FastString -> Module

mkInteractiveModule :: Int -> Module

mkMainModule :: FastString -> Module

mkMainModule_ :: ModuleName -> Module

mkPrimModule :: FastString -> Module

mkRationalBase10IdKey :: Unique

mkRationalBase10Name :: Name

mkRationalBase2IdKey :: Unique

mkRationalBase2Name :: Name

mkThisGhcModule :: FastString -> Module

mkThisGhcModule_ :: ModuleName -> Module

mkTrAppKey :: Unique

mkTrAppName :: Name

mkTrConKey :: Unique

mkTrConName :: Name

mkTrFunKey :: Unique

mkTrFunName :: Name

mkTrTypeKey :: Unique

mkTrTypeName :: Name

mkTyConKey :: Unique

mkUnboundName :: OccName -> Name

mk_known_key_name :: NameSpace -> Module -> FastString -> Unique -> Name

modIntIdKey :: Unique

modIntName :: Name

moduleName_RDR :: RdrName

monadClassKey :: Unique

monadClassName :: Name

monadClass_RDR :: RdrName

monadFailClassKey :: Unique

monadFailClassName :: Name

monadFixClassKey :: Unique

monadFixClassName :: Name

monadPlusClassKey :: Unique

monadPlusClassName :: Name

monoidClassKey :: Unique

monoidClassName :: Name

multMulTyConKey :: Unique

multiplicityTyConKey :: Unique

mutVarPrimTyConKey :: Unique

mutableArrayArrayPrimTyConKey :: Unique

mutableArrayPrimTyConKey :: Unique

mutableByteArrayPrimTyConKey :: Unique

mzipIdKey :: Unique

mzipName :: Name

naturalAddIdKey :: Unique

naturalAddName :: Name

naturalAndIdKey :: Unique

naturalAndName :: Name

naturalAndNotIdKey :: Unique

naturalAndNotName :: Name

naturalBitIdKey :: Unique

naturalBitName :: Name

naturalCompareIdKey :: Unique

naturalCompareName :: Name

naturalEqIdKey :: Unique

naturalEqName :: Name

naturalGcdIdKey :: Unique

naturalGcdName :: Name

naturalGeIdKey :: Unique

naturalGeName :: Name

naturalGtIdKey :: Unique

naturalGtName :: Name

naturalLcmIdKey :: Unique

naturalLcmName :: Name

naturalLeIdKey :: Unique

naturalLeName :: Name

naturalLog2IdKey :: Unique

naturalLog2Name :: Name

naturalLogBaseIdKey :: Unique

naturalLogBaseName :: Name

naturalLogBaseWordIdKey :: Unique

naturalLogBaseWordName :: Name

naturalLtIdKey :: Unique

naturalLtName :: Name

naturalMulIdKey :: Unique

naturalMulName :: Name

naturalNBDataConKey :: Unique

naturalNSDataConKey :: Unique

naturalNeIdKey :: Unique

naturalNeName :: Name

naturalNegateIdKey :: Unique

naturalNegateName :: Name

naturalOrIdKey :: Unique

naturalOrName :: Name

naturalPopCountIdKey :: Unique

naturalPopCountName :: Name

naturalPowModIdKey :: Unique

naturalPowModName :: Name

naturalQuotIdKey :: Unique

naturalQuotName :: Name

naturalQuotRemIdKey :: Unique

naturalQuotRemName :: Name

naturalRemIdKey :: Unique

naturalRemName :: Name

naturalShiftLIdKey :: Unique

naturalShiftLName :: Name

naturalShiftRIdKey :: Unique

naturalShiftRName :: Name

naturalSignumIdKey :: Unique

naturalSignumName :: Name

naturalSizeInBaseIdKey :: Unique

naturalSizeInBaseName :: Name

naturalSubIdKey :: Unique

naturalSubName :: Name

naturalSubThrowIdKey :: Unique

naturalSubThrowName :: Name

naturalSubUnsafeIdKey :: Unique

naturalSubUnsafeName :: Name

naturalTestBitIdKey :: Unique

naturalTestBitName :: Name

naturalToDoubleIdKey :: Unique

naturalToDoubleName :: Name

naturalToFloatIdKey :: Unique

naturalToFloatName :: Name

naturalToWordClampIdKey :: Unique

naturalToWordClampName :: Name

naturalToWordIdKey :: Unique

naturalToWordName :: Name

naturalTyConKey :: Unique

naturalXorIdKey :: Unique

naturalXorName :: Name

negateClassOpKey :: Unique

negateName :: Name

newStablePtrIdKey :: Unique

newStablePtrName :: Name

newStablePtr_RDR :: RdrName

nilDataConKey :: Unique

noMethodBindingErrorIdKey :: Unique

noSelTyConKey :: Unique

noSelTyConName :: Name

noSourceStrictnessDataConKey :: Unique

noSourceStrictnessDataConName :: Name

noSourceUnpackednessDataConKey :: Unique

noSourceUnpackednessDataConName :: Name

noinlineIdKey :: Unique

nonEmptyDataConKey :: Unique

nonEmptyTyConKey :: Unique

nonExhaustiveGuardsErrorIdKey :: Unique

notAssocDataCon_RDR :: RdrName

notAssociativeDataConKey :: Unique

notAssociativeDataConName :: Name

not_RDR :: RdrName

nothingDataConKey :: Unique

nullAddrIdKey :: Unique

null_RDR :: RdrName

numClassKey :: Unique

numClassName :: Name

numClass_RDR :: RdrName

numericClassKeys :: [Unique]

oneDataConKey :: Unique

oneShotKey :: Unique

opaqueTyConKey :: Unique

opaqueTyConName :: Name

ordClassKey :: Unique

ordClassName :: Name

ordClass_RDR :: RdrName

ordEQDataConKey :: Unique

ordEQDataConName :: Name

ordGTDataConKey :: Unique

ordGTDataConName :: Name

ordLTDataConKey :: Unique

ordLTDataConName :: Name

orderingTyConKey :: Unique

orderingTyConName :: Name

otherwiseIdKey :: Unique

otherwiseIdName :: Name

pLUGINS :: Module

pRELUDE :: Module

pRELUDE_NAME :: ModuleName

packageName_RDR :: RdrName

par1DataCon_RDR :: RdrName

par1TyConKey :: Unique

par1TyConName :: Name

parens_RDR :: RdrName

patErrorIdKey :: Unique

pfail_RDR :: RdrName

pluginTyConKey :: Unique

pluginTyConName :: Name

plus_RDR :: RdrName

prec_RDR :: RdrName

pred_RDR :: RdrName

prefixDataCon_RDR :: RdrName

prefixIDataConKey :: Unique

prefixIDataConName :: Name

pretendNameIsInScope :: Name -> Bool

printIdKey :: Unique

printName :: Name

prodDataCon_RDR :: RdrName

prodTyConKey :: Unique

prodTyConName :: Name

proxyHashKey :: Unique

proxyPrimTyConKey :: Unique

ptrTyConKey :: Unique

ptrTyConName :: Name

punc_RDR :: RdrName

pureAClassOpKey :: Unique

pureAName :: Name

pure_RDR :: RdrName

pushCallStackKey :: Unique

pushCallStackName :: Name

r1DataCon_RDR :: RdrName

rANDOM :: Module

rEAD_PREC :: Module

rOOT_MAIN :: Module

rTyConKey :: Unique

rTyConName :: Name

raiseDivZeroIdKey :: Unique

raiseOverflowIdKey :: Unique

raiseUnderflowIdKey :: Unique

randomClassKey :: Unique

randomClassName :: Name

randomGenClassKey :: Unique

randomGenClassName :: Name

range_RDR :: RdrName

ratioDataConKey :: Unique

ratioDataConName :: Name

ratioDataCon_RDR :: RdrName

ratioTyConKey :: Unique

ratioTyConName :: Name

rationalToDoubleIdKey :: Unique

rationalToDoubleName :: Name

rationalToFloatIdKey :: Unique

rationalToFloatName :: Name

rationalTyConKey :: Unique

rationalTyConName :: Name

readClassKey :: Unique

readClassName :: Name

readFieldHash_RDR :: RdrName

readField_RDR :: RdrName

readListDefault_RDR :: RdrName

readListPrecDefault_RDR :: RdrName

readListPrec_RDR :: RdrName

readList_RDR :: RdrName

readPrec_RDR :: RdrName

readSymField_RDR :: RdrName

realClassKey :: Unique

realClassName :: Name

realFloatClassKey :: Unique

realFloatClassName :: Name

realFracClassKey :: Unique

realFracClassName :: Name

realToFracIdKey :: Unique

realToFracName :: Name

realWorldPrimIdKey :: Unique

realWorldTyConKey :: Unique

rec0TyConKey :: Unique

rec0TyConName :: Name

rec1DataCon_RDR :: RdrName

rec1TyConKey :: Unique

rec1TyConName :: Name

recConErrorIdKey :: Unique

recSelErrorIdKey :: Unique

rep1TyConKey :: Unique

rep1TyConName :: Name

repTyConKey :: Unique

repTyConName :: Name

replace_RDR :: RdrName

reset_RDR :: RdrName

returnIOIdKey :: Unique

returnIOName :: Name

returnIO_RDR :: RdrName

returnMClassOpKey :: Unique

returnMName :: Name

returnM_RDR :: RdrName

rightAssocDataCon_RDR :: RdrName

rightAssociativeDataConKey :: Unique

rightAssociativeDataConName :: Name

rightDataConKey :: Unique

rightDataConName :: Name

rightSectionKey :: Unique

right_RDR :: RdrName

rootMainKey :: Unique

runMainIOName :: Name

runMainKey :: Unique

runRWKey :: Unique

runRWName :: Name

runtimeErrorIdKey :: Unique

runtimeRepSimpleDataConKeys :: [Unique]

runtimeRepTyConKey :: Unique

s1TyConKey :: Unique

s1TyConName :: Name

sTyConKey :: Unique

sTyConName :: Name

sYSTEM_IO :: Module

sappendClassOpKey :: Unique

sappendName :: Name

selName_RDR :: RdrName

selectorClassKey :: Unique

selectorClassName :: Name

semigroupClassKey :: Unique

semigroupClassName :: Name

seqIdKey :: Unique

setFieldClassOpKey :: Unique

setFieldName :: Name

showClassKey :: Unique

showClassName :: Name

showCommaSpace_RDR :: RdrName

showParen_RDR :: RdrName

showSpace_RDR :: RdrName

showString_RDR :: RdrName

showsPrec_RDR :: RdrName

shows_RDR :: RdrName

smallArrayPrimTyConKey :: Unique

smallMutableArrayPrimTyConKey :: Unique

sndIdKey :: Unique

sndName :: Name

someTypeRepDataConKey :: Unique

someTypeRepDataConName :: Name

someTypeRepTyConKey :: Unique

someTypeRepTyConName :: Name

sourceLazyDataConKey :: Unique

sourceLazyDataConName :: Name

sourceNoUnpackDataConKey :: Unique

sourceNoUnpackDataConName :: Name

sourceStrictDataConKey :: Unique

sourceStrictDataConName :: Name

sourceUnpackDataConKey :: Unique

sourceUnpackDataConName :: Name

specTyConKey :: Unique

specTyConName :: Name

srcLocDataConKey :: Unique

srcLocDataConName :: Name

stableNameDataConKey :: Unique

stableNamePrimTyConKey :: Unique

stableNameTyConKey :: Unique

stablePtrPrimTyConKey :: Unique

stablePtrTyConKey :: Unique

stablePtrTyConName :: Name

standardClassKeys :: [Unique]

starArrStarArrStarKindRepKey :: Unique

starArrStarArrStarKindRepName :: Name

starArrStarKindRepKey :: Unique

starArrStarKindRepName :: Name

starKindRepKey :: Unique

starKindRepName :: Name

statePrimTyConKey :: Unique

staticPtrDataConKey :: Unique

staticPtrDataConName :: Name

staticPtrInfoDataConKey :: Unique

staticPtrInfoDataConName :: Name

staticPtrInfoTyConKey :: Unique

staticPtrInfoTyConName :: Name

staticPtrTyConKey :: Unique

staticPtrTyConName :: Name

step_RDR :: RdrName

stringTyConKey :: Unique

succ_RDR :: RdrName

sumRepDataConKey :: Unique

sumTyConKey :: Unique

sumTyConName :: Name

symbol_RDR :: RdrName

tVarPrimTyConKey :: Unique

tYPEABLE :: Module

tYPEABLE_INTERNAL :: Module

tYPETyConKey :: Unique

tcQual :: Module -> FastString -> Unique -> Name

tcQual_RDR :: Module -> FastString -> RdrName

thenAClassOpKey :: Unique

thenAName :: Name

thenIOIdKey :: Unique

thenIOName :: Name

thenMClassOpKey :: Unique

thenMName :: Name

threadIdPrimTyConKey :: Unique

times_RDR :: RdrName

to1_RDR :: RdrName

toAnnotationWrapperIdKey :: Unique

toAnnotationWrapperName :: Name

toDynIdKey :: Unique

toDynName :: Name

toEnum_RDR :: RdrName

toIntegerClassOpKey :: Unique

toIntegerName :: Name

toInteger_RDR :: RdrName

toListClassOpKey :: Unique

toListName :: Name

toList_RDR :: RdrName

toRationalClassOpKey :: Unique

toRationalName :: Name

toRational_RDR :: RdrName

to_RDR :: RdrName

tr'PtrRepLiftedKey :: Unique

trGhcPrimModuleKey :: Unique

trGhcPrimModuleName :: Name

trLiftedRepKey :: Unique

trModuleDataConKey :: Unique

trModuleDataConName :: Name

trModuleTyConKey :: Unique

trModuleTyConName :: Name

trNameDDataConKey :: Unique

trNameDDataConName :: Name

trNameSDataConKey :: Unique

trNameSDataConName :: Name

trNameTyConKey :: Unique

trNameTyConName :: Name

trRuntimeRepKey :: Unique

trTYPE'PtrRepLiftedKey :: Unique

trTYPEKey :: Unique

trTyConDataConKey :: Unique

trTyConDataConName :: Name

trTyConTyConKey :: Unique

trTyConTyConName :: Name

traceKey :: Unique

traceName :: Name

traversableClassKey :: Unique

traversableClassName :: Name

traverse_RDR :: RdrName

trueDataConKey :: Unique

tupleRepDataConKey :: Unique

typeCharCmpTyFamNameKey :: Unique

typeCharKindConNameKey :: Unique

typeCharToNatTyFamNameKey :: Unique

typeCharTypeRepKey :: Unique

typeCharTypeRepName :: Name

typeConKey :: Unique

typeConsSymbolTyFamNameKey :: Unique

typeErrorAppendDataConKey :: Unique

typeErrorAppendDataConName :: Name

typeErrorIdKey :: Unique

typeErrorShowTypeDataConKey :: Unique

typeErrorShowTypeDataConName :: Name

typeErrorTextDataConKey :: Unique

typeErrorTextDataConName :: Name

typeErrorVAppendDataConKey :: Unique

typeErrorVAppendDataConName :: Name

typeLeqCharTyFamNameKey :: Unique

typeLitCharDataConKey :: Unique

typeLitCharDataConName :: Name

typeLitNatDataConKey :: Unique

typeLitNatDataConName :: Name

typeLitSortTyConKey :: Unique

typeLitSortTyConName :: Name

typeLitSymbolDataConKey :: Unique

typeLitSymbolDataConName :: Name

typeNatAddTyFamNameKey :: Unique

typeNatCmpTyFamNameKey :: Unique

typeNatDivTyFamNameKey :: Unique

typeNatExpTyFamNameKey :: Unique

typeNatLogTyFamNameKey :: Unique

typeNatModTyFamNameKey :: Unique

typeNatMulTyFamNameKey :: Unique

typeNatSubTyFamNameKey :: Unique

typeNatToCharTyFamNameKey :: Unique

typeNatTypeRepKey :: Unique

typeNatTypeRepName :: Name

typeRepIdKey :: Unique

typeRepIdName :: Name

typeRepTyConKey :: Unique

typeRepTyConName :: Name

typeSymbolAppendFamNameKey :: Unique

typeSymbolCmpTyFamNameKey :: Unique

typeSymbolKindConNameKey :: Unique

typeSymbolTypeRepKey :: Unique

typeSymbolTypeRepName :: Name

typeUnconsSymbolTyFamNameKey :: Unique

typeableClassKey :: Unique

typeableClassName :: Name

u1DataCon_RDR :: RdrName

u1TyConKey :: Unique

u1TyConName :: Name

uAddrDataCon_RDR :: RdrName

uAddrHash_RDR :: RdrName

uAddrTyConKey :: Unique

uAddrTyConName :: Name

uCharDataCon_RDR :: RdrName

uCharHash_RDR :: RdrName

uCharTyConKey :: Unique

uCharTyConName :: Name

uDoubleDataCon_RDR :: RdrName

uDoubleHash_RDR :: RdrName

uDoubleTyConKey :: Unique

uDoubleTyConName :: Name

uFloatDataCon_RDR :: RdrName

uFloatHash_RDR :: RdrName

uFloatTyConKey :: Unique

uFloatTyConName :: Name

uIntDataCon_RDR :: RdrName

uIntHash_RDR :: RdrName

uIntTyConKey :: Unique

uIntTyConName :: Name

uNSAFE_COERCE :: Module

uRecTyConKey :: Unique

uRecTyConName :: Name

uWordDataCon_RDR :: RdrName

uWordHash_RDR :: RdrName

uWordTyConKey :: Unique

uWordTyConName :: Name

unComp1_RDR :: RdrName

unK1_RDR :: RdrName

unPar1_RDR :: RdrName

unRec1_RDR :: RdrName

unboundKey :: Unique

unknown1TyConKey :: Unique

unknown2TyConKey :: Unique

unknown3TyConKey :: Unique

unknownTyConKey :: Unique

unliftedDataConKey :: Unique

unliftedRepTyConKey :: Unique

unliftedTypeKindTyConKey :: Unique

unpackCStringAppendIdKey :: Unique

unpackCStringFoldrIdKey :: Unique

unpackCStringFoldrName :: Name

unpackCStringFoldrUtf8IdKey :: Unique

unpackCStringFoldrUtf8Name :: Name

unpackCStringFoldrUtf8_RDR :: RdrName

unpackCStringFoldr_RDR :: RdrName

unpackCStringIdKey :: Unique

unpackCStringName :: Name

unpackCStringUtf8IdKey :: Unique

unpackCStringUtf8Name :: Name

unpackCStringUtf8_RDR :: RdrName

unpackCString_RDR :: RdrName

unrestrictedFunTyConKey :: Unique

unsafeCoercePrimIdKey :: Unique

unsafeCoercePrimName :: Name

unsafeEqualityProofIdKey :: Unique

unsafeEqualityProofName :: Name

unsafeEqualityTyConKey :: Unique

unsafeEqualityTyConName :: Name

unsafeIndex_RDR :: RdrName

unsafeRangeSize_RDR :: RdrName

unsafeReflDataConKey :: Unique

unsafeReflDataConName :: Name

v1TyConKey :: Unique

v1TyConName :: Name

varQual :: Module -> FastString -> Unique -> Name

varQual_RDR :: Module -> FastString -> RdrName

vecCountDataConKeys :: [Unique]

vecCountTyConKey :: Unique

vecElemDataConKeys :: [Unique]

vecElemTyConKey :: Unique

vecRepDataConKey :: Unique

voidArgIdKey :: Unique

voidPrimIdKey :: Unique

weakPrimTyConKey :: Unique

wildCardKey :: Unique

wildCardName :: Name

word16PrimTyConKey :: Unique

word16TyConKey :: Unique

word16TyConName :: Name

word16X16PrimTyConKey :: Unique

word16X32PrimTyConKey :: Unique

word16X8PrimTyConKey :: Unique

word32PrimTyConKey :: Unique

word32TyConKey :: Unique

word32TyConName :: Name

word32X16PrimTyConKey :: Unique

word32X4PrimTyConKey :: Unique

word32X8PrimTyConKey :: Unique

word64PrimTyConKey :: Unique

word64TyConKey :: Unique

word64TyConName :: Name

word64X2PrimTyConKey :: Unique

word64X4PrimTyConKey :: Unique

word64X8PrimTyConKey :: Unique

word8DataConKey :: Unique

word8PrimTyConKey :: Unique

word8TyConKey :: Unique

word8TyConName :: Name

word8X16PrimTyConKey :: Unique

word8X32PrimTyConKey :: Unique

word8X64PrimTyConKey :: Unique

wordDataConKey :: Unique

wordPrimTyConKey :: Unique

wordTyConKey :: Unique

zipIdKey :: Unique

zipName :: Name