GHC.ByteCode.Types

Description

Bytecode assembler types

Documentation

data CompiledByteCode

Constructors

CompiledByteCode

Fields

bc_bcos :: [UnlinkedBCO]

bc_itbls :: ItblEnv

bc_ffis :: [FFIInfo]

bc_strs :: [RemotePtr ()]

bc_breaks :: Maybe ModBreaks

seqCompiledByteCode :: CompiledByteCode -> ()

newtype FFIInfo

Constructors

FFIInfo (RemotePtr C_ffi_cif)

data UnlinkedBCO

Constructors

UnlinkedBCO

Fields

unlinkedBCOName :: !Name

unlinkedBCOArity :: !Int

unlinkedBCOInstrs :: !(UArray Int Word16)

unlinkedBCOBitmap :: !(UArray Int Word64)

unlinkedBCOLits :: !(SizedSeq BCONPtr)

unlinkedBCOPtrs :: !(SizedSeq BCOPtr)

data BCOPtr

Constructors

BCOPtrName !Name

BCOPtrPrimOp !PrimOp

BCOPtrBCO !UnlinkedBCO

BCOPtrBreakArray

data BCONPtr

Constructors

BCONPtrWord !Word

BCONPtrLbl !FastString

BCONPtrItbl !Name

BCONPtrStr !ByteString

type ItblEnv = NameEnv (Name, ItblPtr)

newtype ItblPtr

Constructors

ItblPtr (RemotePtr StgInfoTable)

data CgBreakInfo

Information about a breakpoint that we know at code-generation time

Constructors

CgBreakInfo

Fields

cgb_vars :: [Maybe (Id, Word16)]

cgb_resty :: Type

data ModBreaks

All the information about the breakpoints for a module

Constructors

ModBreaks

Fields

modBreaks_flags :: ForeignRef BreakArray

The array of flags, one per breakpoint, indicating which breakpoints are enabled.

modBreaks_locs :: !(Array BreakIndex SrcSpan)

An array giving the source span of each breakpoint.

modBreaks_vars :: !(Array BreakIndex [OccName])

An array giving the names of the free variables at each breakpoint.

modBreaks_decls :: !(Array BreakIndex [String])

An array giving the names of the declarations enclosing each breakpoint. See Note [Field modBreaks_decls]

modBreaks_ccs :: !(Array BreakIndex (RemotePtr CostCentre))

Array pointing to cost centre for each breakpoint

modBreaks_breakInfo :: IntMap CgBreakInfo

info about each breakpoint from the bytecode generator

type BreakIndex = Int

Breakpoint index

emptyModBreaks :: ModBreaks

Construct an empty ModBreaks

data CCostCentre

C CostCentre type