GHC.Cmm

Cmm top-level datatypes

type DCmmGroup = GenCmmGroup CmmStatics DCmmTopInfo DCmmGraph

Cmm group after STG generation

type CmmProgram = [CmmGroup]

type CmmGroup = GenCmmGroup CmmStatics CmmTopInfo CmmGraph

Cmm group before SRT generation

type CmmGroupSRTs = GenCmmGroup RawCmmStatics CmmTopInfo CmmGraph

Cmm group with SRTs

type RawCmmGroup = GenCmmGroup RawCmmStatics (LabelMap RawCmmStatics) CmmGraph

Raw cmm group (TODO (osa): not sure what that means)

type GenCmmGroup d h g = [GenCmmDecl d h g]

type CmmDecl = GenCmmDecl CmmStatics CmmTopInfo CmmGraph

type DCmmDecl = GenCmmDecl CmmStatics DCmmTopInfo DCmmGraph

type CmmDeclSRTs = GenCmmDecl RawCmmStatics CmmTopInfo CmmGraph

data GenCmmDecl d h g

A top-level chunk, abstracted over the type of the contents of the basic blocks (Cmm or instructions are the likely instantiations).

Constructors

CmmProc h CLabel [GlobalRegUse] g

CmmData Section d

type CmmDataDecl = GenCmmDataDecl CmmStatics

cmmDataDeclCmmDecl :: GenCmmDataDecl d -> GenCmmDecl d h g

type DCmmGraph = GenGenCmmGraph DWrap CmmNode

type CmmGraph = GenCmmGraph CmmNode

type GenCmmGraph (n :: Extensibility -> Extensibility -> Type) = GenGenCmmGraph LabelMap n

data GenGenCmmGraph (s :: Type -> Type) (n :: Extensibility -> Extensibility -> Type)

Constructors

CmmGraph

Fields

g_entry :: BlockId

g_graph :: Graph' s Block n C C

toBlockMap :: CmmGraph -> LabelMap CmmBlock

revPostorder :: CmmGraph -> [CmmBlock]

toBlockList :: CmmGraph -> [CmmBlock]

type CmmBlock = Block CmmNode C C

type RawCmmDecl = GenCmmDecl RawCmmStatics (LabelMap RawCmmStatics) CmmGraph

data Section

Constructors

Section SectionType CLabel

data SectionType

Constructors

Text

Data

ReadOnlyData

RelocatableReadOnlyData

UninitialisedData

InitArray

FiniArray

CString

OtherSection String

data GenCmmStatics (rawOnly :: Bool) where

Static data before or after SRT generation

Constructors

CmmStatics :: CLabel -> CmmInfoTable -> CostCentreStack -> [CmmLit] -> [CmmLit] -> GenCmmStatics 'False

CmmStaticsRaw :: forall (rawOnly :: Bool). CLabel -> [CmmStatic] -> GenCmmStatics rawOnly
Static data, after SRTs are generated

type CmmStatics = GenCmmStatics 'False

type RawCmmStatics = GenCmmStatics 'True

data CmmStatic

Constructors

CmmStaticLit CmmLit
a literal value, size given by cmmLitRep of the literal.

CmmUninitialised Int
uninitialised data, N bytes long

CmmString ByteString
string of 8-bit values only, not zero terminated.

CmmFileEmbed FilePath Int
an embedded binary file and its byte length

data SectionProtection

Constructors

ReadWriteSection

ReadOnlySection

WriteProtectedSection

sectionProtection :: Section -> SectionProtection

Should a data in this section be considered constant at runtime

newtype DWrap a

Constructors

DWrap [(BlockId, a)]

unDeterm :: DWrap a -> [(BlockId, a)]

removeDeterm :: DCmmGroup -> CmmGroup

removeDetermDecl :: DCmmDecl -> CmmDecl

removeDetermGraph :: DCmmGraph -> CmmGraph

Blocks containing lists

data GenBasicBlock i

Constructors

BasicBlock BlockId [i]

blockId :: GenBasicBlock i -> BlockId

The branch block id is that of the first block in the branch, which is that branch's entry point

newtype ListGraph i

Constructors

ListGraph [GenBasicBlock i]

pprBBlock :: Outputable stmt => GenBasicBlock stmt -> SDoc

Info Tables

data GenCmmTopInfo (f :: Type -> Type)

CmmTopInfo is attached to each CmmDecl (see defn of CmmGroup), and contains the extra info (beyond the executable code) that belongs to that CmmDecl.

Constructors

TopInfo

Fields

info_tbls :: f CmmInfoTable

stack_info :: CmmStackInfo

type DCmmTopInfo = GenCmmTopInfo DWrap

type CmmTopInfo = GenCmmTopInfo LabelMap

data CmmStackInfo

Constructors

StackInfo

Fields

arg_space :: ByteOff

do_layout :: Bool

data CmmInfoTable

Info table as a haskell data type

Constructors

CmmInfoTable

Fields

cit_lbl :: CLabel

cit_rep :: SMRep

cit_prof :: ProfilingInfo

cit_srt :: Maybe CLabel

cit_clo :: Maybe (Id, CostCentreStack)

topInfoTable :: forall a (s :: Type -> Type) (n :: Extensibility -> Extensibility -> Type). GenCmmDecl a CmmTopInfo (GenGenCmmGraph s n) -> Maybe CmmInfoTable

topInfoTableD :: forall a (s :: Type -> Type) (n :: Extensibility -> Extensibility -> Type). GenCmmDecl a DCmmTopInfo (GenGenCmmGraph s n) -> Maybe CmmInfoTable

data ClosureTypeInfo

Constructors

Constr ConTagZ ConstrDescription

Fun FunArity ArgDescr

Thunk

ThunkSelector SelectorOffset

BlackHole

IndStatic

data ProfilingInfo

Constructors

NoProfilingInfo

ProfilingInfo ByteString ByteString

type ConstrDescription = ByteString

Statements, expressions and types

module GHC.Cmm.Node

module GHC.Cmm.Expr

Pretty-printing

pprCmmGroup :: (OutputableP Platform d, OutputableP Platform info, OutputableP Platform g) => Platform -> GenCmmGroup d info g -> SDoc

pprSection :: Platform -> Section -> SDoc

pprStatic :: Platform -> CmmStatic -> SDoc