GHC.Cmm.Dataflow

Documentation

type C = 'Closed

type O = 'Open

data Block n e x

A sequence of nodes. May be any of four shapes (OO, OC, CO, CC). Open at the entry means single entry, mutatis mutandis for exit. A closedclosed block is a basic/ block and can't be extended further. Clients should avoid manipulating blocks and should stick to either nodes or graphs.

lastNode :: Block n x C -> n O C

entryLabel

Arguments

:: NonLocal thing

=> thing C x

-> Label
The label of a first node or block

foldNodesBwdOO :: (CmmNode O O -> f -> f) -> Block CmmNode O O -> f -> f

Folds backward over all nodes of an open-open block. Strict in the accumulator.

foldRewriteNodesBwdOO :: forall f. (CmmNode O O -> f -> UniqSM (Block CmmNode O O, f)) -> Block CmmNode O O -> f -> UniqSM (Block CmmNode O O, f)

Folds backward over all the nodes of an open-open block and allows rewriting them. The accumulator is both the block of nodes and f (usually dataflow facts). Strict in both accumulated parts.

data DataflowLattice a

Constructors

DataflowLattice

Fields

fact_bot :: a

fact_join :: JoinFun a

newtype OldFact a

Constructors

OldFact a

newtype NewFact a

Constructors

NewFact a

data JoinedFact a

The result of joining OldFact and NewFact.

Constructors

Changed !a
Result is different than OldFact.

NotChanged !a
Result is the same as OldFact.

type TransferFun f = CmmBlock -> FactBase f -> FactBase f

type RewriteFun f = CmmBlock -> FactBase f -> UniqSM (CmmBlock, FactBase f)

Function for rewrtiting and analysis combined. To be used with rewriteCmm.

Currently set to work with UniqSM monad, but we could probably abstract that away (if we do that, we might want to specialize the fixpoint algorithms to the particular monads through SPECIALIZE).

type family Fact (x :: Extensibility) f :: Type

type FactBase f = LabelMap f

getFact :: DataflowLattice f -> Label -> FactBase f -> f

mkFactBase :: DataflowLattice f -> [(Label, f)] -> FactBase f

Returns the joined facts for each label.

analyzeCmmFwd :: DataflowLattice f -> TransferFun f -> CmmGraph -> FactBase f -> FactBase f

analyzeCmmBwd :: DataflowLattice f -> TransferFun f -> CmmGraph -> FactBase f -> FactBase f

rewriteCmmBwd :: DataflowLattice f -> RewriteFun f -> CmmGraph -> FactBase f -> UniqSM (CmmGraph, FactBase f)

changedIf :: Bool -> a -> JoinedFact a

joinOutFacts :: NonLocal n => DataflowLattice f -> n e C -> FactBase f -> f

Returns the result of joining the facts from all the successors of the provided node or block.

joinFacts :: DataflowLattice f -> [f] -> f