GHC.Cmm.Expr

Documentation

data CmmExpr

Constructors

CmmLit !CmmLit

CmmLoad !CmmExpr !CmmType !AlignmentSpec

CmmReg !CmmReg

CmmMachOp MachOp [CmmExpr]

CmmStackSlot Area !Int

CmmRegOff !CmmReg !Int

cmmExprType :: Platform -> CmmExpr -> CmmType

cmmExprWidth :: Platform -> CmmExpr -> Width

cmmExprAlignment :: CmmExpr -> Alignment

Returns an alignment in bytes of a CmmExpr when it's a statically known integer constant, otherwise returns an alignment of 1 byte. The caller is responsible for using with a sensible CmmExpr argument.

maybeInvertCmmExpr :: CmmExpr -> Maybe CmmExpr

data CmmReg

Constructors

CmmLocal !LocalReg

CmmGlobal GlobalReg

cmmRegType :: Platform -> CmmReg -> CmmType

cmmRegWidth :: Platform -> CmmReg -> Width

data CmmLit

Constructors

CmmInt !Integer !Width

CmmFloat Rational !Width

CmmVec [CmmLit]

CmmLabel CLabel

CmmLabelOff CLabel !Int

CmmLabelDiffOff CLabel CLabel !Int !Width

CmmBlock !BlockId

CmmHighStackMark

cmmLitType :: Platform -> CmmLit -> CmmType

data AlignmentSpec

Constructors

NaturallyAligned

Unaligned

data LocalReg

Constructors

LocalReg !Unique !CmmType
Parameters: 1. Identifier 2. Type

localRegType :: LocalReg -> CmmType

data GlobalReg

Constructors

VanillaReg !Int VGcPtr

FloatReg !Int

DoubleReg !Int

LongReg !Int

XmmReg !Int

YmmReg !Int

ZmmReg !Int

Sp

SpLim

Hp

HpLim

CCCS

CurrentTSO

CurrentNursery

HpAlloc

EagerBlackholeInfo

GCEnter1

GCFun

BaseReg

MachSp

UnwindReturnReg

PicBaseReg

isArgReg :: GlobalReg -> Bool

globalRegType :: Platform -> GlobalReg -> CmmType

spReg :: CmmReg

hpReg :: CmmReg

spLimReg :: CmmReg

hpLimReg :: CmmReg

nodeReg :: CmmReg

currentTSOReg :: CmmReg

currentNurseryReg :: CmmReg

hpAllocReg :: CmmReg

cccsReg :: CmmReg

node :: GlobalReg

baseReg :: CmmReg

data VGcPtr

Constructors

VGcPtr

VNonGcPtr

class Ord r => DefinerOfRegs r a

Minimal complete definition

foldRegsDefd

class Ord r => UserOfRegs r a

Minimal complete definition

foldRegsUsed

foldRegsDefd :: DefinerOfRegs r a => Platform -> (b -> r -> b) -> b -> a -> b

foldRegsUsed :: UserOfRegs r a => Platform -> (b -> r -> b) -> b -> a -> b

foldLocalRegsDefd :: DefinerOfRegs LocalReg a => Platform -> (b -> LocalReg -> b) -> b -> a -> b

foldLocalRegsUsed :: UserOfRegs LocalReg a => Platform -> (b -> LocalReg -> b) -> b -> a -> b

type RegSet r = Set r

Sets of registers

type LocalRegSet = RegSet LocalReg

type GlobalRegSet = RegSet GlobalReg

emptyRegSet :: RegSet r

elemRegSet :: Ord r => r -> RegSet r -> Bool

extendRegSet :: Ord r => RegSet r -> r -> RegSet r

deleteFromRegSet :: Ord r => RegSet r -> r -> RegSet r

mkRegSet :: Ord r => [r] -> RegSet r

plusRegSet :: Ord r => RegSet r -> RegSet r -> RegSet r

minusRegSet :: Ord r => RegSet r -> RegSet r -> RegSet r

timesRegSet :: Ord r => RegSet r -> RegSet r -> RegSet r

sizeRegSet :: RegSet r -> Int

nullRegSet :: RegSet r -> Bool

regSetToList :: RegSet r -> [r]

data Area

A stack area is either the stack slot where a variable is spilled or the stack space where function arguments and results are passed.

Constructors

Old

Young !BlockId

module GHC.Cmm.MachOp

module GHC.Cmm.Type