GHC.Cmm.UniqueRenamer

Documentation

detRenameCmmGroup :: DetUniqFM -> DCmmGroup -> (DetUniqFM, CmmGroup)

detRenameIPEMap :: DetUniqFM -> InfoTableProvMap -> (DetUniqFM, InfoTableProvMap)

class Monad m => MonadGetUnique (m :: Type -> Type) where

Get a unique from a monad that can access a unique supply.

Crucially, because MonadGetUnique doesn't allow you to get the UniqSupply (unlike MonadUnique), an instance such as UniqDSM can use a deterministic unique supply to return deterministic uniques without allowing for the UniqSupply to be shared.

Methods

getUniqueM :: m Unique

data DetUniqFM

A mapping from non-deterministic uniques to deterministic uniques, to rename local symbols with the end goal of producing deterministic object files. See Note [Renaming uniques deterministically]

emptyDetUFM :: DetUniqFM

module GHC.Types.Unique.DSM