GHC.CmmToAsm.AArch64.Regs

Documentation

allMachRegNos :: [RegNo]

allocatableRegs :: Platform -> [RealReg]

allGpArgRegs :: [Reg]

allFpArgRegs :: [Reg]

data AddrMode

Constructors

AddrRegReg Reg Reg

AddrRegImm Reg Imm

AddrReg Reg

data Imm

Constructors

ImmInt Int

ImmInteger Integer

ImmCLbl CLabel

ImmLit SDoc

ImmIndex CLabel Int

ImmFloat Rational

ImmDouble Rational

ImmConstantSum Imm Imm

ImmConstantDiff Imm Imm

strImmLit :: String -> Imm

litToImm :: CmmLit -> Imm

virtualRegSqueeze :: RegClass -> VirtualReg -> Int

regSqueeze_class reg Calculate the maximum number of register colors that could be denied to a node of this class due to having this reg as a neighbour.

realRegSqueeze :: RegClass -> RealReg -> Int

mkVirtualReg :: Unique -> Format -> VirtualReg

classOfRealReg :: RealReg -> RegClass

regDotColor :: RealReg -> SDoc

Orphan instances

Show SDoc

Instance details

Methods

showsPrec :: Int -> SDoc -> ShowS

show :: SDoc -> String

showList :: [SDoc] -> ShowS

Eq SDoc

Instance details

Methods

(==) :: SDoc -> SDoc -> Bool

(/=) :: SDoc -> SDoc -> Bool