GHC.CmmToAsm.Dwarf.Types

Dwarf information

data DwarfInfo

Individual dwarf records. Each one will be encoded as an entry in the .debug_info section.

Constructors

DwarfCompileUnit

Fields

dwChildren :: [DwarfInfo]

dwName :: String

dwProducer :: String

dwCompDir :: String

dwLowLabel :: CLabel

dwHighLabel :: CLabel

dwLineLabel :: PtrString

DwarfSubprogram

Fields

dwChildren :: [DwarfInfo]

dwName :: String

dwLabel :: CLabel

dwParent :: Maybe CLabel

label of DIE belonging to the parent tick

DwarfBlock

Fields

dwChildren :: [DwarfInfo]

dwLabel :: CLabel

dwMarker :: Maybe CLabel

DwarfSrcNote

Fields

dwSrcSpan :: RealSrcSpan

pprDwarfInfo :: Platform -> Bool -> DwarfInfo -> SDoc

Generate assembly for DWARF data

pprAbbrevDecls :: Platform -> Bool -> SDoc

Abbreviation declaration. This explains the binary encoding we use for representing DwarfInfo. Be aware that this must be updated along with pprDwarfInfo.

Dwarf address range table

data DwarfARange

A DWARF address range. This is used by the debugger to quickly locate which compilation unit a given address belongs to. This type assumes a non-segmented address-space.

Constructors

DwarfARange

Fields

dwArngStartLabel :: CLabel

dwArngEndLabel :: CLabel

pprDwarfARanges :: Platform -> [DwarfARange] -> Unique -> SDoc

Print assembler directives corresponding to a DWARF .debug_aranges address table entry.

Dwarf frame

data DwarfFrame

Information about unwind instructions for a procedure. This corresponds to a "Common Information Entry" (CIE) in DWARF.

Constructors

DwarfFrame

Fields

dwCieLabel :: CLabel

dwCieInit :: UnwindTable

dwCieProcs :: [DwarfFrameProc]

data DwarfFrameProc

Unwind instructions for an individual procedure. Corresponds to a "Frame Description Entry" (FDE) in DWARF.

Constructors

DwarfFrameProc

Fields

dwFdeProc :: CLabel

dwFdeHasInfo :: Bool

dwFdeBlocks :: [DwarfFrameBlock]

List of blocks. Order must match asm!

data DwarfFrameBlock

Unwind instructions for a block. Will become part of the containing FDE.

Constructors

DwarfFrameBlock

Fields

dwFdeBlkHasInfo :: Bool

dwFdeUnwind :: [UnwindPoint]

these unwind points must occur in the same order as they occur in the block

pprDwarfFrame :: Platform -> DwarfFrame -> SDoc

Header for the .debug_frame section. Here we emit the "Common Information Entry" record that establishes general call frame parameters and the default stack layout.

Utilities

pprByte :: Word8 -> SDoc

Assembly for a single byte of constant DWARF data

pprHalf :: Word16 -> SDoc

Assembly for a two-byte constant integer

pprData4' :: SDoc -> SDoc

Assembly for 4 bytes of dynamic DWARF data

pprDwWord :: SDoc -> SDoc

Assembly for a DWARF word of dynamic data. This means 32 bit, as we are generating 32 bit DWARF.

pprWord :: Platform -> SDoc -> SDoc

Assembly for a machine word of dynamic data. Depends on the architecture we are currently generating code for.

pprLEBWord :: Word -> SDoc

Prints a number in "little endian base 128" format. The idea is to optimize for small numbers by stopping once all further bytes would be 0. The highest bit in every byte signals whether there are further bytes to read.

pprLEBInt :: Int -> SDoc

Same as pprLEBWord, but for a signed number

wordAlign :: Platform -> SDoc

Align assembly at (machine) word boundary

sectionOffset :: Platform -> SDoc -> SDoc -> SDoc

Generate an offset into another section. This is tricky because this is handled differently depending on platform: Mac Os expects us to calculate the offset using assembler arithmetic. Linux expects us to just reference the target directly, and will figure out on their own that we actually need an offset. Finally, Windows has a special directive to refer to relative offsets. Fun.