GHC.CmmToAsm.Format

Description

Formats on this architecture A Format is a combination of width and class

TODO: Signed vs unsigned?

TODO: This module is currently shared by all architectures because NCGMonad need to know about it to make a VReg. It would be better to have architecture specific formats, and do the overloading properly. eg SPARC doesn't care about FF80.

Documentation

data Format

Constructors

II8

II16

II32

II64

FF32

FF64

VecFormat

Fields

!Length

number of elements (always at least 2)

!ScalarFormat

format of each element

Bundled Patterns

pattern IntegerFormat :: Format

data ScalarFormat

Constructors

FmtInt8

FmtInt16

FmtInt32

FmtInt64

FmtFloat

FmtDouble

intFormat :: Width -> Format

Get the integer format of this width.

floatFormat :: Width -> Format

Get the float format of this width.

isIntFormat :: Format -> Bool

Check if a format represent an integer value.

isIntScalarFormat :: ScalarFormat -> Bool

intScalarFormat :: Width -> ScalarFormat

isFloatFormat :: Format -> Bool

Check if a format represents a floating point value.

vecFormat :: CmmType -> Format

isVecFormat :: Format -> Bool

Check if a format represents a vector

cmmTypeFormat :: CmmType -> Format

Convert a Cmm type to a Format.

formatToWidth :: Format -> Width

Get the Width of a Format.

scalarWidth :: ScalarFormat -> Width

formatInBytes :: Format -> Int

isFloatScalarFormat :: ScalarFormat -> Bool

isFloatOrFloatVecFormat :: Format -> Bool

floatScalarFormat :: Width -> ScalarFormat

scalarFormatFormat :: ScalarFormat -> Format

data VirtualRegWithFormat

A typed virtual register: a virtual register, together with the specific format we are using it at.

Constructors

VirtualRegWithFormat

Fields

virtualRegWithFormat_reg :: !VirtualReg

virtualRegWithFormat_format :: !Format

data RegWithFormat

A typed register: a register, together with the specific format we are using it at.

Constructors

RegWithFormat

Fields

regWithFormat_reg :: !Reg

regWithFormat_format :: !Format

takeVirtualRegs :: UniqSet RegWithFormat -> UniqSet VirtualReg

Take all the virtual registers from this set.

takeRealRegs :: UniqSet RegWithFormat -> UniqSet RealReg

Take all the real registers from this set.