GHC.CmmToAsm.Monad

Documentation

data NcgImpl statics instr jumpDest

Constructors

NcgImpl

Fields

ncgConfig :: !NCGConfig

cmmTopCodeGen :: RawCmmDecl -> NatM [NatCmmDecl statics instr]

generateJumpTableForInstr :: instr -> Maybe (NatCmmDecl statics instr)

getJumpDestBlockId :: jumpDest -> Maybe BlockId

canShortcut :: instr -> Maybe jumpDest

shortcutStatics :: (BlockId -> Maybe jumpDest) -> statics -> statics

shortcutJump :: (BlockId -> Maybe jumpDest) -> instr -> instr

pprNatCmmDecl :: NatCmmDecl statics instr -> SDoc

Module is only for printing internal labels. See Note [Internal proc labels] in CLabel.

maxSpillSlots :: Int

allocatableRegs :: [RealReg]

ncgExpandTop :: [NatCmmDecl statics instr] -> [NatCmmDecl statics instr]

ncgAllocMoreStack :: Int -> NatCmmDecl statics instr -> UniqSM (NatCmmDecl statics instr, [(BlockId, BlockId)])

The list of block ids records the redirected jumps to allow us to update the CFG.

ncgMakeFarBranches :: LabelMap RawCmmStatics -> [NatBasicBlock instr] -> [NatBasicBlock instr]

extractUnwindPoints :: [instr] -> [UnwindPoint]

given the instruction sequence of a block, produce a list of the block's UnwindPoints See Note [What is this unwinding business?] in GHC.Cmm.DebugBlock and Note [Unwinding information in the NCG] in this module.

invertCondBranches :: Maybe CFG -> LabelMap RawCmmStatics -> [NatBasicBlock instr] -> [NatBasicBlock instr]

Turn the sequence of jcc l1; jmp l2 into jncc l2; <block_l1> when possible.

data NatM_State

Constructors

NatM_State

Fields

natm_us :: UniqSupply

natm_delta :: Int

natm_imports :: [CLabel]

natm_pic :: Maybe Reg

natm_config :: NCGConfig

natm_modloc :: ModLocation

natm_fileid :: DwarfFiles

natm_debug_map :: LabelMap DebugBlock

natm_cfg :: CFG

Having a CFG with additional information is essential for some operations. However we can't reconstruct all information once we generated instructions. So instead we update the CFG as we go.

mkNatM_State :: UniqSupply -> Int -> NCGConfig -> ModLocation -> DwarfFiles -> LabelMap DebugBlock -> CFG -> NatM_State

data NatM result

initNat :: NatM_State -> NatM a -> (a, NatM_State)

addImportNat :: CLabel -> NatM ()

addNodeBetweenNat :: BlockId -> BlockId -> BlockId -> NatM ()

Record that we added a block between from and old.

addImmediateSuccessorNat :: BlockId -> BlockId -> NatM ()

Place succ after block and change any edges block -> X to succ -> X

updateCfgNat :: (CFG -> CFG) -> NatM ()

getUniqueNat :: NatM Unique

mapAccumLNat :: (acc -> x -> NatM (acc, y)) -> acc -> [x] -> NatM (acc, [y])

setDeltaNat :: Int -> NatM ()

getConfig :: NatM NCGConfig

Get native code generator configuration

getPlatform :: NatM Platform

Get target platform from native code generator configuration

getDeltaNat :: NatM Int

getThisModuleNat :: NatM Module

getBlockIdNat :: NatM BlockId

getNewLabelNat :: NatM CLabel

getNewRegNat :: Format -> NatM Reg

getNewRegPairNat :: Format -> NatM (Reg, Reg)

getPicBaseMaybeNat :: NatM (Maybe Reg)

getPicBaseNat :: Format -> NatM Reg

getCfgWeights :: NatM Weights

Get CFG edge weights

getModLoc :: NatM ModLocation

getFileId :: FastString -> NatM Int

getDebugBlock :: Label -> NatM (Maybe DebugBlock)

type DwarfFiles = UniqFM FastString (FastString, Int)