GHC.CmmToAsm.PPC.Instr

Documentation

data Instr

Constructors

COMMENT FastString

LOCATION Int Int Int String

LDATA Section RawCmmStatics

NEWBLOCK BlockId

DELTA Int

LD Format Reg AddrMode

LDFAR Format Reg AddrMode

LDR Format Reg AddrMode

LA Format Reg AddrMode

ST Format Reg AddrMode

STFAR Format Reg AddrMode

STU Format Reg AddrMode

STC Format Reg AddrMode

LIS Reg Imm

LI Reg Imm

MR Reg Reg

CMP Format Reg RI

CMPL Format Reg RI

BCC Cond BlockId (Maybe Bool)

BCCFAR Cond BlockId (Maybe Bool)

JMP CLabel [Reg]

MTCTR Reg

BCTR [Maybe BlockId] (Maybe CLabel) [Reg]

BL CLabel [Reg]

BCTRL [Reg]

ADD Reg Reg RI

ADDO Reg Reg Reg

ADDC Reg Reg Reg

ADDE Reg Reg Reg

ADDZE Reg Reg

ADDIS Reg Reg Imm

SUBF Reg Reg Reg

SUBFO Reg Reg Reg

SUBFC Reg Reg RI

SUBFE Reg Reg Reg

MULL Format Reg Reg RI

MULLO Format Reg Reg Reg

MFOV Format Reg

MULHU Format Reg Reg Reg

DIV Format Bool Reg Reg Reg

AND Reg Reg RI

ANDC Reg Reg Reg

NAND Reg Reg Reg

OR Reg Reg RI

ORIS Reg Reg Imm

XOR Reg Reg RI

XORIS Reg Reg Imm

EXTS Format Reg Reg

CNTLZ Format Reg Reg

NEG Reg Reg

NOT Reg Reg

SL Format Reg Reg RI

SR Format Reg Reg RI

SRA Format Reg Reg RI

RLWINM Reg Reg Int Int Int

CLRLI Format Reg Reg Int

CLRRI Format Reg Reg Int

FADD Format Reg Reg Reg

FSUB Format Reg Reg Reg

FMUL Format Reg Reg Reg

FDIV Format Reg Reg Reg

FABS Reg Reg

FNEG Reg Reg

FCMP Reg Reg

FCTIWZ Reg Reg

FCTIDZ Reg Reg

FCFID Reg Reg

FRSP Reg Reg

CRNOR Int Int Int

MFCR Reg

MFLR Reg

FETCHPC Reg

HWSYNC

ISYNC

LWSYNC

NOP

data RI

Constructors

RIReg Reg

RIImm Imm

archWordFormat :: Bool -> Format

stackFrameHeaderSize :: Platform -> Int

The size of a minimal stackframe header including minimal parameter save area.

maxSpillSlots :: NCGConfig -> Int

The number of spill slots available without allocating more.

allocMoreStack :: Platform -> Int -> NatCmmDecl statics Instr -> UniqSM (NatCmmDecl statics Instr, [(BlockId, BlockId)])

makeFarBranches :: LabelMap RawCmmStatics -> [NatBasicBlock Instr] -> [NatBasicBlock Instr]

mkJumpInstr :: BlockId -> [Instr]

Make an unconditional jump instruction.

mkLoadInstr :: NCGConfig -> Reg -> Int -> Int -> [Instr]

mkSpillInstr :: NCGConfig -> Reg -> Int -> Int -> [Instr]

An instruction to spill a register into a spill slot.

patchJumpInstr :: Instr -> (BlockId -> BlockId) -> Instr

Change the destination of this jump instruction. Used in the linear allocator when adding fixup blocks for join points.

patchRegsOfInstr :: Instr -> (Reg -> Reg) -> Instr

Apply a given mapping to all the register references in this instruction.

jumpDestsOfInstr :: Instr -> [BlockId]

Checks whether this instruction is a jump/branch instruction. One that can change the flow of control in a way that the register allocator needs to worry about.

takeRegRegMoveInstr :: Instr -> Maybe (Reg, Reg)

Take the source and destination from this reg -> reg move instruction or Nothing if it's not one

takeDeltaInstr :: Instr -> Maybe Int

See if this instruction is telling us the current C stack delta

mkRegRegMoveInstr :: Reg -> Reg -> Instr

Copy the value in a register to another one. Must work for all register classes.

mkStackAllocInstr :: Platform -> Int -> [Instr]

mkStackDeallocInstr :: Platform -> Int -> [Instr]

regUsageOfInstr :: Platform -> Instr -> RegUsage

Get the registers that are being used by this instruction. regUsage doesn't need to do any trickery for jumps and such. Just state precisely the regs read and written by that insn. The consequences of control flow transfers, as far as register allocation goes, are taken care of by the register allocator.

isJumpishInstr :: Instr -> Bool

Checks whether this instruction is a jump/branch instruction. One that can change the flow of control in a way that the register allocator needs to worry about.

isMetaInstr :: Instr -> Bool