GHC.CmmToAsm.PPC.Regs

Documentation

virtualRegSqueeze :: RegClass -> VirtualReg -> Int

regSqueeze_class reg Calculate the maximum number of register colors that could be denied to a node of this class due to having this reg as a neighbour.

realRegSqueeze :: RegClass -> RealReg -> Int

mkVirtualReg :: Unique -> Format -> VirtualReg

regDotColor :: RealReg -> SDoc

data Imm

Constructors

ImmInt Int

ImmInteger Integer

ImmCLbl CLabel

ImmLit SDoc

ImmIndex CLabel Int

ImmFloat Rational

ImmDouble Rational

ImmConstantSum Imm Imm

ImmConstantDiff Imm Imm

LO Imm

HI Imm

HA Imm

HIGHERA Imm

HIGHESTA Imm

strImmLit :: String -> Imm

litToImm :: CmmLit -> Imm

data AddrMode

Constructors

AddrRegReg Reg Reg

AddrRegImm Reg Imm

addrOffset :: AddrMode -> Int -> Maybe AddrMode

spRel :: Platform -> Int -> AddrMode

argRegs :: RegNo -> [Reg]

allArgRegs :: [Reg]

callClobberedRegs :: Platform -> [Reg]

allMachRegNos :: [RegNo]

classOfRealReg :: RealReg -> RegClass

showReg :: RegNo -> String

toRegNo :: Reg -> RegNo

allFPArgRegs :: Platform -> [Reg]

fits16Bits :: Integral a => a -> Bool

makeImmediate :: Integral a => Width -> Bool -> a -> Maybe Imm

fReg :: Int -> RegNo

r0 :: Reg

sp :: Reg

toc :: Reg

r3 :: Reg

r4 :: Reg

r11 :: Reg

r12 :: Reg

r30 :: Reg

tmpReg :: Platform -> Reg

f1 :: Reg

allocatableRegs :: Platform -> [RealReg]