GHC.CmmToAsm.RV64.Regs

Registers

x0RegNo :: RegNo

First integer register number. zero register.

x1RegNo :: RegNo

return address register

raRegNo :: RegNo

return address register

x5RegNo :: RegNo

t0RegNo :: RegNo

x7RegNo :: RegNo

t2RegNo :: RegNo

x28RegNo :: RegNo

t3RegNo :: RegNo

x31RegNo :: RegNo

Last integer register number. Used as TMP (IP) register.

t6RegNo :: RegNo

Last integer register number. Used as TMP (IP) register.

tmpRegNo :: RegNo

Last integer register number. Used as TMP (IP) register.

d0RegNo :: RegNo

First floating point register.

ft0RegNo :: RegNo

First floating point register.

d7RegNo :: RegNo

ft7RegNo :: RegNo

d31RegNo :: RegNo

Last floating point register.

x10RegNo :: RegNo

a0RegNo :: RegNo

x17RegNo :: RegNo

a7RegNo :: RegNo

d10RegNo :: RegNo

fa0RegNo :: RegNo

d17RegNo :: RegNo

fa7RegNo :: RegNo

zeroReg :: Reg

raReg :: Reg

spMachReg :: Reg

Not to be confused with the CmmReg spReg

tmpReg :: Reg

allMachRegNos :: [RegNo]

All machine register numbers.

allocatableRegs :: Platform -> [RealReg]

Registers available to the register allocator.

These are all registers minus those with a fixed role in RISCV ABI (zero, lr, sp, gp, tp, fp, tmp) and GHC RTS (Base, Sp, Hp, HpLim, R1..R8, F1..F6, D1..D6.)

allGpArgRegs :: [Reg]

Integer argument registers according to the calling convention

allFpArgRegs :: [Reg]

Floating point argument registers according to the calling convention

Addressing modes

data AddrMode

Addressing modes

Constructors

AddrRegImm Reg Imm
A register plus some immediate integer, e.g. 8(sp) or -16(sp). The offset needs to fit into 12bits.

AddrReg Reg
A register

Immediates

data Imm

Constructors

ImmInt Int

ImmInteger Integer

ImmCLbl CLabel

ImmLit FastString

ImmIndex CLabel Int

ImmFloat Rational

ImmDouble Rational

ImmConstantSum Imm Imm

ImmConstantDiff Imm Imm

litToImm :: CmmLit -> Imm

Map CmmLit to Imm

N.B. this is a partial function, because not all CmmLits have an immediate representation.

virtualRegSqueeze :: RegClass -> VirtualReg -> Int

regSqueeze_class reg Calculate the maximum number of register colors that could be denied to a node of this class due to having this reg as a neighbour.

realRegSqueeze :: RegClass -> RealReg -> Int

mkVirtualReg :: Unique -> Format -> VirtualReg

classOfRealReg :: RealReg -> RegClass

regDotColor :: RealReg -> SDoc