GHC.CmmToAsm.Reg.Graph.SpillCost

Documentation

type SpillCostRecord = (VirtualReg, Int, Int, Int)

Records the expected cost to spill some register.

plusSpillCostRecord :: SpillCostRecord -> SpillCostRecord -> SpillCostRecord

Add two spill cost records.

pprSpillCostRecord :: (VirtualReg -> RegClass) -> (Reg -> SDoc) -> Graph VirtualReg RegClass RealReg -> SpillCostRecord -> SDoc

Show a spill cost record, including the degree from the graph and final calculated spill cost.

type SpillCostInfo = UniqFM VirtualReg SpillCostRecord

Map of SpillCostRecord

zeroSpillCostInfo :: SpillCostInfo

An empty map of spill costs.

plusSpillCostInfo :: SpillCostInfo -> SpillCostInfo -> SpillCostInfo

Add two spill cost infos.

slurpSpillCostInfo :: forall instr statics. Instruction instr => Platform -> Maybe CFG -> LiveCmmDecl statics instr -> SpillCostInfo

Slurp out information used for determining spill costs.

For each vreg, the number of times it was written to, read from, and the number of instructions it was live on entry to (lifetime)

chooseSpill :: SpillCostInfo -> Graph VirtualReg RegClass RealReg -> VirtualReg

Choose a node to spill from this graph

lifeMapFromSpillCostInfo :: SpillCostInfo -> UniqFM VirtualReg (VirtualReg, Int)

Extract a map of register lifetimes from a SpillCostInfo.