GHC.CmmToAsm.Reg.Linear.StackMap

Description

The assignment of virtual registers to stack slots

Documentation

type StackSlot = Int

Identifier for a stack slot.

data StackMap

Constructors

StackMap

Fields

stackMapNextFreeSlot :: !Int

The slots that are still available to be allocated.

stackMapAssignment :: UniqFM Unique StackSlot

Assignment of vregs to stack slots.

emptyStackMap :: StackMap

An empty stack map, with all slots available.

getStackSlotFor :: StackMap -> Unique -> (StackMap, Int)

If this vreg unique already has a stack assignment then return the slot number, otherwise allocate a new slot, and update the map.

getStackUse :: StackMap -> Int

Return the number of stack slots that were allocated