GHC.CmmToAsm.Reg.Linear.State

Description

State monad for the linear register allocator.

Documentation

data RA_State freeRegs

The register allocator state

Constructors

RA_State

Fields

ra_blockassig :: BlockAssignment freeRegs

the current mapping from basic blocks to the register assignments at the beginning of that block.

ra_freeregs :: !freeRegs

free machine registers

ra_assig :: RegMap Loc

assignment of temps to locations

ra_delta :: Int

current stack delta

ra_stack :: StackMap

free stack slots for spilling

ra_us :: UniqSupply

unique supply for generating names for join point fixup blocks.

ra_spills :: [SpillReason]

Record why things were spilled, for -ddrop-asm-stats. Just keep a list here instead of a map of regs -> reasons. We don't want to slow down the allocator if we're not going to emit the stats.

ra_config :: !NCGConfig

Native code generator configuration

ra_fixups :: [(BlockId, BlockId, BlockId)]

(from,fixup,to) : We inserted fixup code between from and to

data RegM freeRegs a

The register allocator monad type.

runR :: NCGConfig -> BlockAssignment freeRegs -> freeRegs -> RegMap Loc -> StackMap -> UniqSupply -> RegM freeRegs a -> (BlockAssignment freeRegs, StackMap, RegAllocStats, a)

Run a computation in the RegM register allocator monad.

spillR :: Instruction instr => Reg -> Unique -> RegM freeRegs ([instr], Int)

loadR :: Instruction instr => Reg -> Int -> RegM freeRegs [instr]

getFreeRegsR :: RegM freeRegs freeRegs

setFreeRegsR :: freeRegs -> RegM freeRegs ()

getAssigR :: RegM freeRegs (RegMap Loc)

setAssigR :: RegMap Loc -> RegM freeRegs ()

getBlockAssigR :: RegM freeRegs (BlockAssignment freeRegs)

setBlockAssigR :: BlockAssignment freeRegs -> RegM freeRegs ()

setDeltaR :: Int -> RegM freeRegs ()

getDeltaR :: RegM freeRegs Int

getUniqueR :: RegM freeRegs Unique

getConfig :: RegM a NCGConfig

Get native code generator configuration

getPlatform :: RegM a Platform

Get target platform from native code generator configuration

recordSpill :: SpillReason -> RegM freeRegs ()

Record that a spill instruction was inserted, for profiling.

recordFixupBlock :: BlockId -> BlockId -> BlockId -> RegM freeRegs ()

Record a created fixup block