GHC.CmmToAsm.Types

Documentation

type NatCmm instr = GenCmmGroup RawCmmStatics (LabelMap RawCmmStatics) (ListGraph instr)

type NatCmmDecl statics instr = GenCmmDecl statics (LabelMap RawCmmStatics) (ListGraph instr)

type NatBasicBlock instr = GenBasicBlock instr

data GenBasicBlock i

Constructors

BasicBlock BlockId [i]

blockId :: GenBasicBlock i -> BlockId

The branch block id is that of the first block in the branch, which is that branch's entry point

newtype ListGraph i

Constructors

ListGraph [GenBasicBlock i]

type RawCmmStatics = GenCmmStatics 'True

type RawCmmDecl = GenCmmDecl RawCmmStatics (LabelMap RawCmmStatics) CmmGraph