GHC.CmmToAsm.X86.Instr

Documentation

data Instr

Constructors

COMMENT FastString

LOCATION Int Int Int String

LDATA Section (Alignment, RawCmmStatics)

NEWBLOCK BlockId

UNWIND CLabel UnwindTable

DELTA Int

MOV Format Operand Operand

CMOV Cond Format Operand Reg

MOVZxL Format Operand Operand

MOVSxL Format Operand Operand

LEA Format Operand Operand

ADD Format Operand Operand

ADC Format Operand Operand

SUB Format Operand Operand

SBB Format Operand Operand

MUL Format Operand Operand

MUL2 Format Operand

IMUL Format Operand Operand

IMUL2 Format Operand

DIV Format Operand

IDIV Format Operand

ADD_CC Format Operand Operand

SUB_CC Format Operand Operand

AND Format Operand Operand

OR Format Operand Operand

XOR Format Operand Operand

NOT Format Operand

NEGI Format Operand

BSWAP Format Reg

SHL Format Operand Operand

SAR Format Operand Operand

SHR Format Operand Operand

BT Format Imm Operand

NOP

X87Store Format AddrMode

CVTSS2SD Reg Reg

CVTSD2SS Reg Reg

CVTTSS2SIQ Format Operand Reg

CVTTSD2SIQ Format Operand Reg

CVTSI2SS Format Operand Reg

CVTSI2SD Format Operand Reg

FDIV Format Operand Operand

SQRT Format Operand Reg

TEST Format Operand Operand

CMP Format Operand Operand

SETCC Cond Operand

PUSH Format Operand

POP Format Operand

JMP Operand [Reg]

JXX Cond BlockId

JXX_GBL Cond Imm

JMP_TBL Operand [Maybe JumpDest] Section CLabel

CALL
X86 call instruction

Fields

(Either Imm Reg)

Jump target

[Reg]

Arguments (required for register allocation)

CLTD Format

FETCHGOT Reg

FETCHPC Reg

POPCNT Format Operand Reg

LZCNT Format Operand Reg

TZCNT Format Operand Reg

BSF Format Operand Reg

BSR Format Operand Reg

PDEP Format Operand Operand Reg

PEXT Format Operand Operand Reg

PREFETCH PrefetchVariant Format Operand

LOCK Instr

XADD Format Operand Operand

CMPXCHG Format Operand Operand

XCHG Format Operand Reg

MFENCE

data Operand

Constructors

OpReg Reg

OpImm Imm

OpAddr AddrMode

data PrefetchVariant

Constructors

NTA

Lvl0

Lvl1

Lvl2

data JumpDest

Constructors

DestBlockId BlockId

DestImm Imm

getJumpDestBlockId :: JumpDest -> Maybe BlockId

canShortcut :: Instr -> Maybe JumpDest

shortcutStatics :: (BlockId -> Maybe JumpDest) -> (Alignment, RawCmmStatics) -> (Alignment, RawCmmStatics)

shortcutJump :: (BlockId -> Maybe JumpDest) -> Instr -> Instr

allocMoreStack :: Platform -> Int -> NatCmmDecl statics Instr -> UniqSM (NatCmmDecl statics Instr, [(BlockId, BlockId)])

maxSpillSlots :: NCGConfig -> Int

archWordFormat :: Bool -> Format

takeRegRegMoveInstr :: Instr -> Maybe (Reg, Reg)

Check whether an instruction represents a reg-reg move. The register allocator attempts to eliminate reg->reg moves whenever it can, by assigning the src and dest temporaries to the same real register.

regUsageOfInstr :: Platform -> Instr -> RegUsage

Returns which registers are read and written as a (read, written) pair.

takeDeltaInstr :: Instr -> Maybe Int

See if this instruction is telling us the current C stack delta

mkLoadInstr :: NCGConfig -> Reg -> Int -> Int -> [Instr]

Make a spill reload instruction.

mkJumpInstr :: BlockId -> [Instr]

Make an unconditional branch instruction.

mkStackAllocInstr :: Platform -> Int -> [Instr]

mkStackDeallocInstr :: Platform -> Int -> [Instr]

mkSpillInstr :: NCGConfig -> Reg -> Int -> Int -> [Instr]

Make a spill instruction.

mkRegRegMoveInstr :: Platform -> Reg -> Reg -> Instr

Make a reg-reg move instruction. On SPARC v8 there are no instructions to move directly between floating point and integer regs. If we need to do that then we have to go via memory.

jumpDestsOfInstr :: Instr -> [BlockId]

patchRegsOfInstr :: Instr -> (Reg -> Reg) -> Instr

Applies the supplied function to all registers in instructions. Typically used to change virtual registers to real registers.

patchJumpInstr :: Instr -> (BlockId -> BlockId) -> Instr

isMetaInstr :: Instr -> Bool

isJumpishInstr :: Instr -> Bool