GHC.CmmToLlvm.Base

Description

Base LLVM Code Generation module

Contains functions useful through out the code generator.

Documentation

type LlvmCmmDecl = GenCmmDecl [LlvmData] (Maybe RawCmmStatics) (ListGraph LlvmStatement)

type LlvmBasicBlock = GenBasicBlock LlvmStatement

type LiveGlobalRegs = [GlobalReg]

Global registers live on proc entry

type LlvmUnresData = (CLabel, Section, LlvmType, [UnresStatic])

Unresolved code. Of the form: (data label, data type, unresolved data)

type LlvmData = ([LMGlobal], [LlvmType])

Top level LLVM Data (globals and type aliases)

type UnresLabel = CmmLit

An unresolved Label.

Labels are unresolved when we haven't yet determined if they are defined in the module we are currently compiling, or an external one.

type UnresStatic = Either UnresLabel LlvmStatic

data LlvmVersion

supportedLlvmVersionLowerBound :: LlvmVersion

The (inclusive) lower bound on the LLVM Version that is currently supported.

supportedLlvmVersionUpperBound :: LlvmVersion

The (not-inclusive) upper bound bound on the LLVM Version that is currently supported.

llvmVersionSupported :: LlvmVersion -> Bool

parseLlvmVersion :: String -> Maybe LlvmVersion

llvmVersionStr :: LlvmVersion -> String

llvmVersionList :: LlvmVersion -> [Int]

data LlvmM a

The Llvm monad. Wraps LlvmEnv state as well as the IO monad

runLlvm :: Logger -> DynFlags -> LlvmVersion -> BufHandle -> LlvmM a -> IO a

Get initial Llvm environment.

withClearVars :: LlvmM a -> LlvmM a

Clear variables from the environment for a subcomputation

varLookup :: Uniquable key => key -> LlvmM (Maybe LlvmType)

Lookup variables or functions in the environment.

varInsert :: Uniquable key => key -> LlvmType -> LlvmM ()

Insert variables or functions into the environment.

markStackReg :: GlobalReg -> LlvmM ()

Set a register as allocated on the stack

checkStackReg :: GlobalReg -> LlvmM Bool

Check whether a register is allocated on the stack

funLookup :: Uniquable key => key -> LlvmM (Maybe LlvmType)

Lookup variables or functions in the environment.

funInsert :: Uniquable key => key -> LlvmType -> LlvmM ()

Insert variables or functions into the environment.

getLlvmVer :: LlvmM LlvmVersion

Get the LLVM version we are generating code for

getDynFlags :: HasDynFlags m => m DynFlags

dumpIfSetLlvm :: DumpFlag -> String -> DumpFormat -> SDoc -> LlvmM ()

Dumps the document if the corresponding flag has been set by the user

renderLlvm :: SDoc -> LlvmM ()

Prints the given contents to the output handle

markUsedVar :: LlvmVar -> LlvmM ()

Marks a variable as "used"

getUsedVars :: LlvmM [LlvmVar]

Return all variables marked as "used" so far

ghcInternalFunctions :: LlvmM ()

Here we pre-initialise some functions that are used internally by GHC so as to make sure they have the most general type in the case that user code also uses these functions but with a different type than GHC internally. (Main offender is treating return type as void instead of 'void *'). Fixes trac #5486.

getPlatform :: LlvmM Platform

Get target platform

getLlvmOpts :: LlvmM LlvmOpts

Get LLVM options

getMetaUniqueId :: LlvmM MetaId

Allocate a new global unnamed metadata identifier

setUniqMeta :: Unique -> MetaId -> LlvmM ()

Sets metadata node for a given unique

getUniqMeta :: Unique -> LlvmM (Maybe MetaId)

Gets metadata node for given unique

liftIO :: IO a -> LlvmM a

Lifting of IO actions. Not exported, as we want to encapsulate IO.

cmmToLlvmType :: CmmType -> LlvmType

Translate a basic CmmType to an LlvmType.

widthToLlvmFloat :: Width -> LlvmType

Translate a Cmm Float Width to a LlvmType.

widthToLlvmInt :: Width -> LlvmType

Translate a Cmm Bit Width to a LlvmType.

llvmFunTy :: LiveGlobalRegs -> LlvmM LlvmType

Llvm Function type for Cmm function

llvmFunSig :: LiveGlobalRegs -> CLabel -> LlvmLinkageType -> LlvmM LlvmFunctionDecl

Llvm Function signature

llvmFunArgs :: Platform -> LiveGlobalRegs -> [LlvmVar]

A Function's arguments

llvmStdFunAttrs :: [LlvmFuncAttr]

Llvm standard fun attributes

llvmFunAlign :: Platform -> LMAlign

Alignment to use for functions

llvmInfAlign :: Platform -> LMAlign

Alignment to use for into tables

llvmPtrBits :: Platform -> Int

Pointer width

tysToParams :: [LlvmType] -> [LlvmParameter]

Convert a list of types to a list of function parameters (each with no parameter attributes)

llvmFunSection :: LlvmOpts -> LMString -> LMSection

Section to use for a function

padLiveArgs :: Platform -> LiveGlobalRegs -> LiveGlobalRegs

Return a list of "padding" registers for LLVM function calls.

When we generate LLVM function signatures, we can't just make any register alive on function entry. Instead, we need to insert fake arguments of the same register class until we are sure that one of them is mapped to the register we want alive. E.g. to ensure that F5 is alive, we may need to insert fake arguments mapped to F1, F2, F3 and F4.

Invariant: Cmm FPR regs with number "n" maps to real registers with number "n" If the calling convention uses registers in a different order or if the invariant doesn't hold, this code probably won't be correct.

isFPR :: GlobalReg -> Bool

strCLabel_llvm :: CLabel -> LlvmM LMString

Pretty print a CLabel.

getGlobalPtr :: LMString -> LlvmM LlvmVar

Create/get a pointer to a global value. Might return an alias if the value in question hasn't been defined yet. We especially make no guarantees on the type of the returned pointer.

generateExternDecls :: LlvmM ([LMGlobal], [LlvmType])

Generate definitions for aliases forward-referenced by getGlobalPtr.

Must be called at a point where we are sure that no new global definitions will be generated anymore!

aliasify :: LMGlobal -> LlvmM [LMGlobal]

Here we take a global variable definition, rename it with a $def suffix, and generate the appropriate alias.

llvmDefLabel :: LMString -> LMString

Derive the definition label. It has an identified structure type.