GHC.CmmToLlvm.Mangler

Description

GHC LLVM Mangler

This script processes the assembly produced by LLVM, rewriting all symbols of type function to object. This keeps them from going through the PLT, which would be bad due to tables-next-to-code. On x86_64, it also rewrites AVX instructions that require alignment to their unaligned counterparts, since the stack is only 16-byte aligned but these instructions require 32-byte alignment.

Documentation

llvmFixupAsm :: Logger -> DynFlags -> FilePath -> FilePath -> IO ()

Read in assembly file and process