GHC.CmmToLlvm.Ppr

Description

Pretty print helpers for the LLVM Code generator.

Documentation

pprLlvmCmmDecl :: LlvmCmmDecl -> LlvmM (SDoc, [LlvmVar])

Pretty print LLVM code

pprLlvmData :: LlvmOpts -> LlvmData -> SDoc

Pretty print LLVM data code

infoSection :: String

The section we are putting info tables and their entry code into, should be unique since we process the assembly pattern matching this.