GHC.Core.Coercion.Axiom

Description

Module for coercion axioms, used to represent type family instances and newtypes

Documentation

data BranchFlag

type Branched = 'Branched

type Unbranched = 'Unbranched

type BranchIndex = Int

newtype Branches (br :: BranchFlag)

Constructors

MkBranches

Fields

unMkBranches :: Array BranchIndex CoAxBranch

manyBranches :: [CoAxBranch] -> Branches Branched

unbranched :: CoAxBranch -> Branches Unbranched

fromBranches :: forall (br :: BranchFlag). Branches br -> [CoAxBranch]

numBranches :: forall (br :: BranchFlag). Branches br -> Int

mapAccumBranches :: forall (br :: BranchFlag). ([CoAxBranch] -> CoAxBranch -> CoAxBranch) -> Branches br -> Branches br

The [CoAxBranch] passed into the mapping function is a list of all previous branches, reversed

data CoAxiom (br :: BranchFlag)

A CoAxiom is a "coercion constructor", i.e. a named equality axiom.

Constructors

CoAxiom

Fields

co_ax_unique :: Unique

co_ax_name :: Name

co_ax_role :: Role

co_ax_tc :: TyCon

co_ax_branches :: Branches br

co_ax_implicit :: Bool

data CoAxBranch

A branch of a coercion axiom, which provides the evidence for unwrapping a newtype or a type-family reduction step using a single equation.

Constructors

CoAxBranch

Fields

cab_loc :: SrcSpan

Location of the defining equation See Note [CoAxiom locations]

cab_tvs :: [TyVar]

Bound type variables; not necessarily fresh See Note [CoAxBranch type variables]

cab_eta_tvs :: [TyVar]

Eta-reduced tyvars cab_tvs and cab_lhs may be eta-reduced; see Note [Eta reduction for data families]

cab_cvs :: [CoVar]

Bound coercion variables Always empty, for now. See Note [Constraints in patterns] in GHC.Tc.TyCl

cab_roles :: [Role]

See Note [CoAxBranch roles]

cab_lhs :: [Type]

Type patterns to match against

cab_rhs :: Type

Right-hand side of the equality See Note [CoAxioms are homogeneous]

cab_incomps :: [CoAxBranch]

The previous incompatible branches See Note [Storing compatibility]

toBranchedAxiom :: forall (br :: BranchFlag). CoAxiom br -> CoAxiom Branched

toUnbranchedAxiom :: forall (br :: BranchFlag). CoAxiom br -> CoAxiom Unbranched

coAxiomName :: forall (br :: BranchFlag). CoAxiom br -> Name

coAxiomArity :: forall (br :: BranchFlag). CoAxiom br -> BranchIndex -> Arity

coAxiomBranches :: forall (br :: BranchFlag). CoAxiom br -> Branches br

coAxiomTyCon :: forall (br :: BranchFlag). CoAxiom br -> TyCon

isImplicitCoAxiom :: forall (br :: BranchFlag). CoAxiom br -> Bool

coAxiomNumPats :: forall (br :: BranchFlag). CoAxiom br -> Int

coAxiomNthBranch :: forall (br :: BranchFlag). CoAxiom br -> BranchIndex -> CoAxBranch

coAxiomSingleBranch_maybe :: forall (br :: BranchFlag). CoAxiom br -> Maybe CoAxBranch

coAxiomRole :: forall (br :: BranchFlag). CoAxiom br -> Role

coAxiomSingleBranch :: CoAxiom Unbranched -> CoAxBranch

coAxBranchTyVars :: CoAxBranch -> [TyVar]

coAxBranchCoVars :: CoAxBranch -> [CoVar]

coAxBranchRoles :: CoAxBranch -> [Role]

coAxBranchLHS :: CoAxBranch -> [Type]

coAxBranchRHS :: CoAxBranch -> Type

coAxBranchSpan :: CoAxBranch -> SrcSpan

coAxBranchIncomps :: CoAxBranch -> [CoAxBranch]

placeHolderIncomps :: [CoAxBranch]

data Role

See Note [Roles] in GHC.Core.Coercion

Order of constructors matters: the Ord instance coincides with the *super*typing relation on roles.

Constructors

Nominal

Representational

Phantom

fsFromRole :: Role -> FastString

data CoAxiomRule

CoAxiomRule describes a built-in axiom, one that we assume to be true See Note [CoAxiomRule]

Constructors

BuiltInFamRew BuiltInFamRewrite

BuiltInFamInj BuiltInFamInjectivity

BranchedAxiom (CoAxiom Branched) BranchIndex

UnbranchedAxiom (CoAxiom Unbranched)

data BuiltInFamRewrite

Constructors

BIF_Rewrite

Fields

bifrw_name :: FastString

bifrw_axr :: CoAxiomRule

bifrw_fam_tc :: TyCon

bifrw_arity :: Arity

bifrw_match :: [Type] -> Maybe ([Type], Type)

bifrw_proves :: [TypeEqn] -> Maybe TypeEqn

data BuiltInFamInjectivity

Constructors

BIF_Interact

Fields

bifinj_name :: FastString

bifinj_axr :: CoAxiomRule

bifinj_proves :: TypeEqn -> Maybe TypeEqn

Always unary: just one TypeEqn argument Returns Nothing when it doesn't like the supplied argument. When this happens in a coercion that means that the coercion is ill-formed, and Core Lint checks for that.

type TypeEqn = Pair Type

A more explicit representation for `t1 ~ t2`.

coAxiomRuleArgRoles :: CoAxiomRule -> [Role]

coAxiomRuleRole :: CoAxiomRule -> Role

coAxiomRuleBranch_maybe :: CoAxiomRule -> Maybe (TyCon, Role, CoAxBranch)

isNewtypeAxiomRule_maybe :: CoAxiomRule -> Maybe (TyCon, CoAxBranch)

data BuiltInSynFamily

Constructors

BuiltInSynFamily

Fields

sfMatchFam :: [BuiltInFamRewrite]

sfInteract :: [BuiltInFamInjectivity]

trivialBuiltInFamily :: BuiltInSynFamily

Orphan instances

Binary Role

Instance details

Methods

put_ :: WriteBinHandle -> Role -> IO ()

put :: WriteBinHandle -> Role -> IO (Bin Role)

get :: ReadBinHandle -> IO Role

Outputable Role

Instance details

Methods

ppr :: Role -> SDoc