GHC.Core.ConLike

Documentation

data ConLike

A constructor-like thing

Constructors

RealDataCon DataCon

PatSynCon PatSyn

isVanillaConLike :: ConLike -> Bool

Is this a 'vanilla' constructor-like thing (no existentials, no provided constraints)?

conLikeArity :: ConLike -> Arity

Number of arguments

conLikeFieldLabels :: ConLike -> [FieldLabel]

Names of fields used for selectors

conLikeInstOrigArgTys :: ConLike -> [Type] -> [Scaled Type]

Returns just the instantiated value argument types of a ConLike, (excluding dictionary args)

conLikeUserTyVarBinders :: ConLike -> [InvisTVBinder]

TyVarBinders for the type variables of the ConLike. For pattern synonyms, this will always consist of the universally quantified variables followed by the existentially quantified type variables. For data constructors, the situation is slightly more complicated—see Note [DataCon user type variable binders] in GHC.Core.DataCon.

conLikeExTyCoVars :: ConLike -> [TyCoVar]

Existentially quantified type/coercion variables

conLikeName :: ConLike -> Name

conLikeStupidTheta :: ConLike -> ThetaType

The "stupid theta" of the ConLike, such as data Eq a in:

data Eq a => T a = ...

It is empty for PatSynCon as they do not allow such contexts.

conLikeImplBangs :: ConLike -> [HsImplBang]

Returns the strictness information for each constructor

conLikeFullSig :: ConLike -> ([TyVar], [TyCoVar], [EqSpec], ThetaType, ThetaType, [Scaled Type], Type)

The "full signature" of the ConLike returns, in order:

1) The universally quantified type variables

2) The existentially quantified type/coercion variables

3) The equality specification

4) The provided theta (the constraints provided by a match)

5) The required theta (the constraints required for a match)

6) The original argument types (i.e. before any change of the representation of the type)

7) The original result type

conLikeResTy :: ConLike -> [Type] -> Type

Returns the type of the whole pattern

conLikeFieldType :: ConLike -> FieldLabelString -> Type

Extract the type for any given labelled field of the ConLike

conLikesWithFields :: [ConLike] -> [FieldLabelString] -> [ConLike]

The ConLikes that have *all* the given fields

conLikeIsInfix :: ConLike -> Bool

conLikeHasBuilder :: ConLike -> Bool

conLikeHasBuilder returns True except for uni-directional pattern synonyms, which have no builder