GHC.Core.DataCon

Main data types

data DataCon

A data constructor

AnnKeywordId : AnnOpen, AnnClose,AnnComma

data DataConRep

Data Constructor Representation See Note [Data constructor workers and wrappers]

Constructors

NoDataConRep

DCR

Fields

dcr_wrap_id :: Id

dcr_boxer :: DataConBoxer

dcr_arg_tys :: [Scaled Type]

dcr_stricts :: [StrictnessMark]

dcr_bangs :: [HsImplBang]

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie ~

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{-# UNPACK #-} specified

SrcNoUnpack
{-# NOUNPACK #-} specified

NoSrcUnpack
no unpack pragma

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as the user wrote them in the source code.

(HsSrcBang _ SrcUnpack SrcLazy) and (HsSrcBang _ SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsSrcBang _ NoSrcUnpack SrcLazy)

Constructors

HsSrcBang SourceText SrcUnpackedness SrcStrictness

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field, or one with an unlifted type

HsStrict
Strict but not unpacked field

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data StrictnessMark

Constructors

MarkedStrict

NotMarkedStrict

type ConTag = Int

A *one-index* constructor tag

Type of the tags associated with each constructor possibility or superclass selector

Equality specs

data EqSpec

An EqSpec is a tyvar/type pair representing an equality made in rejigging a GADT constructor

mkEqSpec :: TyVar -> Type -> EqSpec

Make a non-dependent EqSpec

eqSpecTyVar :: EqSpec -> TyVar

eqSpecType :: EqSpec -> Type

eqSpecPair :: EqSpec -> (TyVar, Type)

eqSpecPreds :: [EqSpec] -> ThetaType

substEqSpec :: TCvSubst -> EqSpec -> EqSpec

Substitute in an EqSpec. Precondition: if the LHS of the EqSpec is mapped in the substitution, it is mapped to a type variable, not a full type.

filterEqSpec :: [EqSpec] -> [TyVar] -> [TyVar]

Filter out any TyVars mentioned in an EqSpec.

Field labels

data FieldLabel

Fields in an algebraic record type; see Note [FieldLabel].

Constructors

FieldLabel

Fields

flLabel :: FieldLabelString

User-visible label of the field

flHasDuplicateRecordFields :: DuplicateRecordFields

Was DuplicateRecordFields on in the defining module for this datatype?

flHasFieldSelector :: FieldSelectors

Was FieldSelectors enabled in the defining module for this datatype? See Note [NoFieldSelectors] in GHC.Rename.Env

flSelector :: Name

Record selector function

type FieldLabelString = FastString

Field labels are just represented as strings; they are not necessarily unique (even within a module)

Type construction

mkDataCon

Arguments

:: Name

-> Bool
Is the constructor declared infix?

-> TyConRepName
TyConRepName for the promoted TyCon

-> [HsSrcBang]
Strictness/unpack annotations, from user

-> [FieldLabel]
Field labels for the constructor, if it is a record, otherwise empty

-> [TyVar]
Universals.

-> [TyCoVar]
Existentials.

-> [InvisTVBinder]
User-written TyVarBinders. These must be Inferred/Specified. See Note [TyVarBinders in DataCons]

-> [EqSpec]
GADT equalities

-> KnotTied ThetaType
Theta-type occurring before the arguments proper

-> [KnotTied (Scaled Type)]
Original argument types

-> KnotTied Type
Original result type

-> RuntimeRepInfo
See comments on RuntimeRepInfo

-> KnotTied TyCon
Representation type constructor

-> ConTag
Constructor tag

-> ThetaType
The "stupid theta", context of the data declaration e.g. data Eq a => T a ...

-> Id
Worker Id

-> DataConRep
Representation

-> DataCon

Build a new data constructor

fIRST_TAG :: ConTag

Tags are allocated from here for real constructors or for superclass selectors

Type deconstruction

dataConRepType :: DataCon -> Type

The representation type of the data constructor, i.e. the sort type that will represent values of this type at runtime

dataConInstSig :: DataCon -> [Type] -> ([TyCoVar], ThetaType, [Type])

Instantiate the universal tyvars of a data con, returning ( instantiated existentials , instantiated constraints including dependent GADT equalities which are *also* listed in the instantiated existentials , instantiated args)

dataConFullSig :: DataCon -> ([TyVar], [TyCoVar], [EqSpec], ThetaType, [Scaled Type], Type)

The "full signature" of the DataCon returns, in order:

1) The result of dataConUnivTyVars

2) The result of dataConExTyCoVars

3) The non-dependent GADT equalities. Dependent GADT equalities are implied by coercion variables in return value (2).

4) The other constraints of the data constructor type, excluding GADT equalities

5) The original argument types to the DataCon (i.e. before any change of the representation of the type) with linearity annotations

6) The original result type of the DataCon

dataConName :: DataCon -> Name

The Name of the DataCon, giving it a unique, rooted identification

dataConIdentity :: DataCon -> ByteString

The string package:module.name identifying a constructor, which is attached to its info table and used by the GHCi debugger and the heap profiler

dataConTag :: DataCon -> ConTag

The tag used for ordering DataCons

dataConTagZ :: DataCon -> ConTagZ

dataConTyCon :: DataCon -> TyCon

The type constructor that we are building via this data constructor

dataConOrigTyCon :: DataCon -> TyCon

The original type constructor used in the definition of this data constructor. In case of a data family instance, that will be the family type constructor.

dataConWrapperType :: DataCon -> Type

The user-declared type of the data constructor in the nice-to-read form:

T :: forall a b. a -> b -> T [a]

rather than:

T :: forall a c. forall b. (c~[a]) => a -> b -> T c

The type variables are quantified in the order that the user wrote them. See Note [DataCon user type variable binders].

NB: If the constructor is part of a data instance, the result type mentions the family tycon, not the internal one.

dataConNonlinearType :: DataCon -> Type

dataConDisplayType :: Bool -> DataCon -> Type

dataConUnivTyVars :: DataCon -> [TyVar]

The universally-quantified type variables of the constructor

dataConExTyCoVars :: DataCon -> [TyCoVar]

The existentially-quantified type/coercion variables of the constructor including dependent (kind-) GADT equalities

dataConUnivAndExTyCoVars :: DataCon -> [TyCoVar]

Both the universal and existential type/coercion variables of the constructor

dataConUserTyVars :: DataCon -> [TyVar]

The type variables of the constructor, in the order the user wrote them

dataConUserTyVarBinders :: DataCon -> [InvisTVBinder]

InvisTVBinders for the type variables of the constructor, in the order the user wrote them

dataConEqSpec :: DataCon -> [EqSpec]

Equalities derived from the result type of the data constructor, as written by the programmer in any GADT declaration. This includes *all* GADT-like equalities, including those written in by hand by the programmer.

dataConTheta :: DataCon -> ThetaType

The *full* constraints on the constructor type, including dependent GADT equalities.

dataConStupidTheta :: DataCon -> ThetaType

The "stupid theta" of the DataCon, such as data Eq a in:

data Eq a => T a = ...

dataConOtherTheta :: DataCon -> ThetaType

Returns constraints in the wrapper type, other than those in the dataConEqSpec

dataConInstArgTys

Arguments

:: DataCon
A datacon with no existentials or equality constraints However, it can have a dcTheta (notably it can be a class dictionary, with superclasses)

-> [Type]
Instantiated at these types

-> [Scaled Type]

Finds the instantiated types of the arguments required to construct a DataCon representation NB: these INCLUDE any dictionary args but EXCLUDE the data-declaration context, which is discarded It's all post-flattening etc; this is a representation type

dataConOrigArgTys :: DataCon -> [Scaled Type]

Returns the argument types of the wrapper, excluding all dictionary arguments and without substituting for any type variables

dataConOrigResTy :: DataCon -> Type

dataConInstOrigArgTys :: DataCon -> [Type] -> [Scaled Type]

Returns just the instantiated value argument types of a DataCon, (excluding dictionary args)

dataConRepArgTys :: DataCon -> [Scaled Type]

Returns the arg types of the worker, including *all* non-dependent evidence, after any flattening has been done and without substituting for any type variables

dataConFieldLabels :: DataCon -> [FieldLabel]

The labels for the fields of this particular DataCon

dataConFieldType :: DataCon -> FieldLabelString -> Type

Extract the type for any given labelled field of the DataCon

dataConFieldType_maybe :: DataCon -> FieldLabelString -> Maybe (FieldLabel, Type)

Extract the label and type for any given labelled field of the DataCon, or return Nothing if the field does not belong to it

dataConSrcBangs :: DataCon -> [HsSrcBang]

Strictness/unpack annotations, from user; or, for imported DataCons, from the interface file The list is in one-to-one correspondence with the arity of the DataCon

dataConSourceArity :: DataCon -> Arity

Source-level arity of the data constructor

dataConRepArity :: DataCon -> Arity

Gives the number of actual fields in the representation of the data constructor. This may be more than appear in the source code; the extra ones are the existentially quantified dictionaries

dataConIsInfix :: DataCon -> Bool

Should the DataCon be presented infix?

dataConWorkId :: DataCon -> Id

Get the Id of the DataCon worker: a function that is the "actual" constructor and has no top level binding in the program. The type may be different from the obvious one written in the source program. Panics if there is no such Id for this DataCon

dataConWrapId :: DataCon -> Id

Returns an Id which looks like the Haskell-source constructor by using the wrapper if it exists (see dataConWrapId_maybe) and failing over to the worker (see dataConWorkId)

dataConWrapId_maybe :: DataCon -> Maybe Id

Get the Id of the DataCon wrapper: a function that wraps the "actual" constructor so it has the type visible in the source program: c.f. dataConWorkId. Returns Nothing if there is no wrapper, which occurs for an algebraic data constructor and also for a newtype (whose constructor is inlined compulsorily)

dataConImplicitTyThings :: DataCon -> [TyThing]

Find all the Ids implicitly brought into scope by the data constructor. Currently, the union of the dataConWorkId and the dataConWrapId

dataConRepStrictness :: DataCon -> [StrictnessMark]

Give the demands on the arguments of a Core constructor application (Con dc args)

dataConImplBangs :: DataCon -> [HsImplBang]

dataConBoxer :: DataCon -> Maybe DataConBoxer

splitDataProductType_maybe

Arguments

:: Type
A product type, perhaps

-> Maybe (TyCon, [Type], DataCon, [Scaled Type])

Extract the type constructor, type argument, data constructor and it's representation argument types from a type if it is a product type.

Precisely, we return Just for any data type that is all of:

Concrete (i.e. constructors visible)

Single-constructor

... which has no existentials

Whether the type is a data type or a newtype.

Predicates on DataCons

isNullarySrcDataCon :: DataCon -> Bool

Return whether there are any argument types for this DataCons original source type See Note [DataCon arities]

isNullaryRepDataCon :: DataCon -> Bool

Return whether there are any argument types for this DataCons runtime representation type See Note [DataCon arities]

isTupleDataCon :: DataCon -> Bool

isBoxedTupleDataCon :: DataCon -> Bool

isUnboxedTupleDataCon :: DataCon -> Bool

isUnboxedSumDataCon :: DataCon -> Bool

isVanillaDataCon :: DataCon -> Bool

Vanilla DataCons are those that are nice boring Haskell 98 constructors

isNewDataCon :: DataCon -> Bool

Is this the DataCon of a newtype?

classDataCon :: Class -> DataCon

dataConCannotMatch :: [Type] -> DataCon -> Bool

dataConUserTyVarsArePermuted :: DataCon -> Bool

Were the type variables of the data con written in a different order than the regular order (universal tyvars followed by existential tyvars)?

This is not a cheap test, so we minimize its use in GHC as much as possible. Currently, its only call site in the GHC codebase is in mkDataConRep in MkId, and so dataConUserTyVarsArePermuted is only called at most once during a data constructor's lifetime.

isBanged :: HsImplBang -> Bool

isMarkedStrict :: StrictnessMark -> Bool

eqHsBang :: HsImplBang -> HsImplBang -> Bool

Compare strictness annotations

isSrcStrict :: SrcStrictness -> Bool

isSrcUnpacked :: SrcUnpackedness -> Bool

specialPromotedDc :: DataCon -> Bool

Should this DataCon be allowed in a type even without -XDataKinds? Currently, only Lifted & Unlifted

Promotion related functions

promoteDataCon :: DataCon -> TyCon