GHC.Core.LateCC.OverloadedCalls

Documentation

overloadedCallsCC :: CoreBind -> LateCCM OverloadedCallsCCState CoreBind

Insert cost centres on function applications with dictionary arguments. The source locations attached to the cost centres is approximated based on the "closest" source note encountered in the traversal.