GHC.Core.LateCC.TopLevelBinds

Documentation

topLevelBindsCCMG :: ModGuts -> CoreM ModGuts

Add late cost centres directly to the ModGuts. This is used inside the core pipeline with the -fprof-late-inline flag. It should not be used after tidy, since it does not manually track inserted cost centers. See Note [Collecting late cost centres].

topLevelBindsCC :: (CoreExpr -> Bool) -> CoreBind -> LateCCM s CoreBind

Insert cost centres on top-level bindings in the module, depending on whether or not they satisfy the given predicate.