GHC.Core.LateCC.Types

Description

Types related to late cost center insertion

Documentation

data LateCCConfig

Late cost center insertion configuration.

Specifies whether cost centers are added to overloaded function call sites and/or top-level bindings, and which top-level bindings they are added to. Also holds the cost center insertion environment.

Constructors

LateCCConfig

Fields

lateCCConfig_whichBinds :: !LateCCBindSpec

lateCCConfig_overloadedCalls :: !Bool

lateCCConfig_env :: !LateCCEnv

data LateCCBindSpec

The types of top-level bindings we support adding cost centers to.

Constructors

LateCCNone

LateCCBinds

LateCCOverloadedBinds

data LateCCEnv

Late cost centre insertion environment

Constructors

LateCCEnv

Fields

lateCCEnv_module :: !Module

Current module

lateCCEnv_file :: Maybe FastString

Current file, if we have one

lateCCEnv_countEntries :: !Bool

Whether the inserted cost centers should count entries

lateCCEnv_collectCCs :: !Bool

Whether to collect the cost centres we insert. See Note [Collecting late cost centres]

data LateCCState s

Late cost centre insertion state, indexed by some extra state type that an insertion method may require.

Constructors

LateCCState

Fields

lateCCState_ccs :: !(Set CostCentre)

Cost centres that have been inserted

lateCCState_ccState :: !CostCentreState

Per-module state tracking for cost centre indices

lateCCState_extra :: !s

initLateCCState :: s -> LateCCState s

The empty late cost centre insertion state

type LateCCM s = ReaderT LateCCEnv (State (LateCCState s))

Late cost centre insertion monad