GHC.Core.Map.Type

Re-export generic interface

class TrieMap m where

Associated Types

type Key m :: Type

Methods

emptyTM :: m a

lookupTM :: forall b. Key m -> m b -> Maybe b

alterTM :: forall b. Key m -> XT b -> m b -> m b

mapTM :: (a -> b) -> m a -> m b

filterTM :: (a -> Bool) -> m a -> m a

foldTM :: (a -> b -> b) -> m a -> b -> b

type XT a = Maybe a -> Maybe a

Maps over Types

data TypeMap a

TypeMap a is a map from Type to a. If you are a client, this is the type you want. The keys in this map may have different kinds.

emptyTypeMap :: TypeMap a

extendTypeMap :: TypeMap a -> Type -> a -> TypeMap a

lookupTypeMap :: TypeMap a -> Type -> Maybe a

foldTypeMap :: (a -> b -> b) -> b -> TypeMap a -> b

data LooseTypeMap a

A LooseTypeMap doesn't do a kind-check. Thus, when lookup up (t |> g), you'll find entries inserted under (t), even if (g) is non-reflexive.

With explicit scoping

data CmEnv

lookupCME :: CmEnv -> Var -> Maybe BoundVar

extendTypeMapWithScope :: TypeMap a -> CmEnv -> Type -> a -> TypeMap a

Extend a TypeMap with a type in the given context. extendTypeMapWithScope m (mkDeBruijnContext [a,b,c]) t v is equivalent to extendTypeMap m (forall a b c. t) v, but allows reuse of the context over multiple insertions.

lookupTypeMapWithScope :: TypeMap a -> CmEnv -> Type -> Maybe a

mkDeBruijnContext :: [Var] -> CmEnv

Construct a deBruijn environment with the given variables in scope. e.g. mkDeBruijnEnv [a,b,c] constructs a context forall a b c.

extendCME :: CmEnv -> Var -> CmEnv

extendCMEs :: CmEnv -> [Var] -> CmEnv

emptyCME :: CmEnv

Utilities for use by friends only

type TypeMapG = GenMap TypeMapX

TypeMapG a is a map from DeBruijn Type to a. The extended key makes it suitable for recursive traversal, since it can track binders, but it is strictly internal to this module. If you are including a TypeMap inside another TrieMap, this is the type you want. Note that this lookup does not do a kind-check. Thus, all keys in this map must have the same kind. Also note that this map respects the distinction between Type and Constraint, despite the fact that they are equivalent type synonyms in Core.

type CoercionMapG = GenMap CoercionMapX

data DeBruijn a

DeBruijn a represents a modulo alpha-renaming. This is achieved by equipping the value with a CmEnv, which tracks an on-the-fly deBruijn numbering. This allows us to define an Eq instance for DeBruijn a, even if this was not (easily) possible for a. Note: we purposely don't export the constructor. Make a helper function if you find yourself needing it.

Constructors

D CmEnv a

deBruijnize :: a -> DeBruijn a

Synthesizes a DeBruijn a from an a, by assuming that there are no bound binders (an empty CmEnv). This is usually what you want if there isn't already a CmEnv in scope.

data BndrMap a

A BndrMap is a TypeMapG which allows us to distinguish between binding forms whose binders have different types. For example, if we are doing a TrieMap lookup on (x :: Int) -> (), we should not pick up an entry in the TrieMap for (x :: Bool) -> (): we can disambiguate this by matching on the type (or kind, if this a binder in a type) of the binder.

We also need to do the same for multiplicity! Which, since multiplicities are encoded simply as a Type, amounts to have a Trie for a pair of types. Tries of pairs are composition.

xtBndr :: forall a. CmEnv -> Var -> XT a -> BndrMap a -> BndrMap a

lkBndr :: CmEnv -> Var -> BndrMap a -> Maybe a

data VarMap a

xtVar :: CmEnv -> Var -> XT a -> VarMap a -> VarMap a

lkVar :: CmEnv -> Var -> VarMap a -> Maybe a

lkDFreeVar :: Var -> DVarEnv a -> Maybe a

xtDFreeVar :: Var -> XT a -> DVarEnv a -> DVarEnv a

xtDNamed :: NamedThing n => n -> XT a -> DNameEnv a -> DNameEnv a

lkDNamed :: NamedThing n => n -> DNameEnv a -> Maybe a