GHC.Core.Opt.Arity

Description

Arity and eta expansion

Documentation

manifestArity :: CoreExpr -> Arity

manifestArity sees how many leading value lambdas there are, after looking through casts

joinRhsArity :: CoreExpr -> JoinArity

exprArity :: CoreExpr -> Arity

An approximate, fast, version of exprEtaExpandArity

typeArity :: Type -> [OneShotInfo]

exprEtaExpandArity :: DynFlags -> CoreExpr -> ArityType

findRhsArity :: DynFlags -> Id -> CoreExpr -> Arity -> ArityType

etaExpand :: Arity -> CoreExpr -> CoreExpr

etaExpand n e returns an expression with the same meaning as e, but with arity n.

Given:

e' = etaExpand n e

We should have that:

ty = exprType e = exprType e'

etaExpandAT :: ArityType -> CoreExpr -> CoreExpr

etaExpandToJoinPoint :: JoinArity -> CoreExpr -> ([CoreBndr], CoreExpr)

Split an expression into the given number of binders and a body, eta-expanding if necessary. Counts value *and* type binders.

etaExpandToJoinPointRule :: JoinArity -> CoreRule -> CoreRule

exprBotStrictness_maybe :: CoreExpr -> Maybe (Arity, StrictSig)

data ArityType

Constructors

ATop [OneShotInfo]

ABot Arity

expandableArityType :: ArityType -> Bool

arityTypeArity :: ArityType -> Arity

maxWithArity :: ArityType -> Arity -> ArityType

isBotArityType :: ArityType -> Bool

idArityType :: Id -> ArityType