GHC.Core.Opt.DmdAnal

Documentation

newtype DmdAnalOpts

Options for the demand analysis

Constructors

DmdAnalOpts

Fields

dmd_strict_dicts :: Bool

Use strict dictionaries

dmdAnalProgram :: DmdAnalOpts -> FamInstEnvs -> [CoreRule] -> CoreProgram -> CoreProgram

Outputs a new copy of the Core program in which binders have been annotated with demand and strictness information.

Note: use seqBinds on the result to avoid leaks due to lazyness (cf Note [Stamp out space leaks in demand analysis])