GHC.Core.Opt.Simplify

Documentation

data SimplifyExprOpts

Configuration record for simplifyExpr. The values of this datatype are only driven by the demands of that function.

Constructors

SimplifyExprOpts

Fields

se_fam_inst :: ![FamInst]

se_mode :: !SimplMode

se_top_env_cfg :: !TopEnvConfig

data SimplifyOpts

Configuration record for simplifyPgm. The values of this datatype are only driven by the demands of that function.

Constructors

SimplifyOpts

Fields

so_dump_core_sizes :: !Bool

so_iterations :: !Int

so_mode :: !SimplMode

so_pass_result_cfg :: !(Maybe LintPassResultConfig)

so_hpt_rules :: !RuleBase

so_top_env_cfg :: !TopEnvConfig

simplifyExpr :: Logger -> ExternalUnitCache -> SimplifyExprOpts -> CoreExpr -> IO CoreExpr

simplifyPgm :: Logger -> UnitEnv -> NamePprCtx -> SimplifyOpts -> ModGuts -> IO (SimplCount, ModGuts)