GHC.Core.Opt.Simplify.Monad

Documentation

data TopEnvConfig

Constructors

TopEnvConfig

Fields

te_history_size :: !Int

te_tick_factor :: !Int

data SimplM result

initSmpl

Arguments

:: Logger

-> IO RuleEnv

-> TopEnvConfig

-> Int
Size of the bindings, used to limit the number of ticks we allow

-> SimplM a

-> IO (a, SimplCount)

traceSmpl :: String -> SDoc -> SimplM ()

getSimplRules :: SimplM RuleEnv

class Monad m => MonadUnique m where

A monad for generating unique identifiers

Minimal complete definition

getUniqueSupplyM

Methods

getUniqueSupplyM :: m UniqSupply

Get a new UniqueSupply

getUniqueM :: m Unique

Get a new unique identifier

getUniquesM :: m [Unique]

Get an infinite list of new unique identifiers

newId :: FastString -> Mult -> Type -> SimplM Id

newJoinId :: [Var] -> Type -> SimplM Id

Make a join id with given type and arity but without call-by-value annotations.

data SimplCount

tick :: Tick -> SimplM ()

freeTick :: Tick -> SimplM ()

checkedTick :: Tick -> SimplM ()

getSimplCount :: SimplM SimplCount

zeroSimplCount

Arguments

:: Bool

ddump-simpl-stats

-> SimplCount

pprSimplCount :: SimplCount -> SDoc

plusSimplCount :: SimplCount -> SimplCount -> SimplCount

isZeroSimplCount :: SimplCount -> Bool