GHC.Core.RoughMap

Description

RoughMap is an approximate finite map data structure keyed on [RoughMatchTc]. This is useful when keying maps on lists of Types (e.g. an instance head).

RoughMatchTc

data RoughMatchTc

Constructors

RM_KnownTc Name

RM_WildCard

isRoughWildcard :: RoughMatchTc -> Bool

typeToRoughMatchTc :: Type -> RoughMatchTc

data RoughMatchLookupTc

Constructors

RML_KnownTc Name
The position only matches the specified KnownTc

RML_NoKnownTc
The position definitely doesn't match any KnownTc

RML_WildCard
The position can match anything

typeToRoughMatchLookupTc :: Type -> RoughMatchLookupTc

roughMatchTcToLookup :: RoughMatchTc -> RoughMatchLookupTc

roughMatchTcs :: [Type] -> [RoughMatchTc]

roughMatchTcsLookup :: [Type] -> [RoughMatchLookupTc]

instanceCantMatch :: [RoughMatchTc] -> [RoughMatchTc] -> Bool

RoughMap

data RoughMap a

Trie of [RoughMatchTc]

Examples* insert [OtherTc] 1 insert [OtherTc] 2 lookup [OtherTc] == [1,2]

emptyRM :: RoughMap a

lookupRM :: [RoughMatchLookupTc] -> RoughMap a -> [a]

Order of result is deterministic.

lookupRM' :: [RoughMatchLookupTc] -> RoughMap a -> (Bag a, [a])

N.B. Returns a Bag for matches, which allows us to avoid rebuilding all of the lists we find in rm_empty, which would otherwise be necessary due to ++ if we returned a list. We use a list for unifiers because the tail is computed lazily and we often only care about the first couple of potential unifiers. Constructing a bag forces the tail which performs much too much work.

See Note [Matching a RoughMap] See Note [Matches vs Unifiers]

insertRM :: [RoughMatchTc] -> a -> RoughMap a -> RoughMap a

filterRM :: (a -> Bool) -> RoughMap a -> RoughMap a

filterMatchingRM :: (a -> Bool) -> [RoughMatchTc] -> RoughMap a -> RoughMap a

Filter all elements that might match a particular key with the given predicate.

elemsRM :: RoughMap a -> [a]

sizeRM :: RoughMap a -> Int

foldRM :: (a -> b -> b) -> b -> RoughMap a -> b

unionRM :: RoughMap a -> RoughMap a -> RoughMap a