GHC.Core.TyCo.Rep

Types

data Type

Constructors

TyVarTy Var
Vanilla type or kind variable (*never* a coercion variable)

AppTy Type Type
Type application to something other than a TyCon. Parameters:

1) Function: must not be a TyConApp or CastTy, must be another AppTy, or TyVarTy See Note [Respecting definitional equality] (EQ1) about the no CastTy requirement

2) Argument type

TyConApp TyCon [KindOrType]
Application of a TyCon, including newtypes and synonyms. Invariant: saturated applications of FunTyCon must use FunTy and saturated synonyms must use their own constructors. However, unsaturated FunTyCons do appear as TyConApps. Parameters:

1) Type constructor being applied to.

2) Type arguments. Might not have enough type arguments here to saturate the constructor. Even type synonyms are not necessarily saturated; for example unsaturated type synonyms can appear as the right hand side of a type synonym.

ForAllTy !ForAllTyBinder Type
A Π type. Note [When we quantify over a coercion variable] INVARIANT: If the binder is a coercion variable, it must be mentioned in the Type. See Note [Unused coercion variable in ForAllTy]

FunTy
FUN m t1 t2 Very common, so an important special case See Note [Function types]

Fields

ft_af :: FunTyFlag

ft_mult :: Mult

ft_arg :: Type

ft_res :: Type

LitTy TyLit
Type literals are similar to type constructors.

CastTy Type KindCoercion
A kind cast. The coercion is always nominal. INVARIANT: The cast is never reflexive (EQ2) INVARIANT: The Type is not a CastTy (use TransCo instead) (EQ3) INVARIANT: The Type is not a ForAllTy over a tyvar (EQ4) See Note [Respecting definitional equality]

CoercionTy Coercion
Injection of a Coercion into a type This should only ever be used in the RHS of an AppTy, in the list of a TyConApp, when applying a promoted GADT data constructor

data TyLit

Constructors

NumTyLit Integer

StrTyLit FastString

CharTyLit Char

type KindOrType = Type

The key representation of types within the compiler

type Kind = Type

The key type representing kinds in the compiler.

type RuntimeRepType = Type

Type synonym used for types of kind RuntimeRep.

type LevityType = Type

Type synonym used for types of kind Levity.

type KnotTied ty = ty

A type labeled KnotTied might have knot-tied tycons in it. See Note [Type checking recursive type and class declarations] in GHC.Tc.TyCl

type PredType = Type

A type of the form p of constraint kind represents a value whose type is the Haskell predicate p, where a predicate is what occurs before the => in a Haskell type.

We use PredType as documentation to mark those types that we guarantee to have this kind.

It can be expanded into its representation, but:

The type checker must treat it as opaque

The rest of the compiler treats it as transparent

Consider these examples:

f :: (Eq a) => a -> Int
g :: (?x :: Int -> Int) => a -> Int
h :: (r\l) => {r} => {l::Int | r}

Here the Eq a and ?x :: Int -> Int and rl are all called "predicates"

type ThetaType = [PredType]

A collection of PredTypes

type FRRType = Type

data ForAllTyFlag

ForAllTyFlag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, ForAllTyBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible Specificity

Required

Bundled Patterns

pattern Specified :: ForAllTyFlag

pattern Inferred :: ForAllTyFlag

data FunTyFlag

The non-dependent version of ForAllTyFlag. See Note [FunTyFlag] Appears here partly so that it's together with its friends ForAllTyFlag and ForallVisFlag, but also because it is used in IfaceType, rather early in the compilation chain

Constructors

FTF_T_T

FTF_T_C

FTF_C_T

FTF_C_C

Coercions

data Coercion

A Coercion is concrete evidence of the equality/convertibility of two types.

Constructors

Refl Type

GRefl Role Type MCoercionN

TyConAppCo Role TyCon [Coercion]

AppCo Coercion CoercionN

ForAllCo TyCoVar KindCoercion Coercion

FunCo

Fields

fco_role :: Role

fco_afl :: FunTyFlag

fco_afr :: FunTyFlag

fco_mult :: CoercionN

fco_arg, fco_res :: Coercion

CoVarCo CoVar

AxiomInstCo (CoAxiom Branched) BranchIndex [Coercion]

AxiomRuleCo CoAxiomRule [Coercion]

UnivCo UnivCoProvenance Role Type Type

SymCo Coercion

TransCo Coercion Coercion

SelCo CoSel Coercion

LRCo LeftOrRight CoercionN

InstCo Coercion CoercionN

KindCo Coercion

SubCo CoercionN

HoleCo CoercionHole
See Note [Coercion holes] Only present during typechecking

data CoSel

Constructors

SelTyCon Int Role

SelFun FunSel

SelForAll

data FunSel

Constructors

SelMult

SelArg

SelRes

data UnivCoProvenance

For simplicity, we have just one UnivCo that represents a coercion from some type to some other type, with (in general) no restrictions on the type. The UnivCoProvenance specifies more exactly what the coercion really is and why a program should (or shouldn't!) trust the coercion. It is reasonable to consider each constructor of UnivCoProvenance as a totally independent coercion form; their only commonality is that they don't tell you what types they coercion between. (That info is in the UnivCo constructor of Coercion.

Constructors

PhantomProv KindCoercion
See Note [Phantom coercions]. Only in Phantom roled coercions

ProofIrrelProv KindCoercion
From the fact that any two coercions are considered equivalent. See Note [ProofIrrelProv]. Can be used in Nominal or Representational coercions

PluginProv String
From a plugin, which asserts that this coercion is sound. The string is for the use of the plugin.

CorePrepProv Bool

data CoercionHole

A coercion to be filled in by the type-checker. See Note [Coercion holes]

Constructors

CoercionHole

Fields

ch_co_var :: CoVar

ch_ref :: IORef (Maybe Coercion)

coHoleCoVar :: CoercionHole -> CoVar

setCoHoleCoVar :: CoercionHole -> CoVar -> CoercionHole

type CoercionN = Coercion

type CoercionR = Coercion

type CoercionP = Coercion

type KindCoercion = CoercionN

data MCoercion

A semantically more meaningful type to represent what may or may not be a useful Coercion.

Constructors

MRefl

MCo Coercion

type MCoercionR = MCoercion

type MCoercionN = MCoercion

Functions over types

mkNakedTyConTy :: TyCon -> Type

mkNakedTyConTy creates a nullary TyConApp. In general you should rather use mkTyConTy, which picks the shared nullary TyConApp from inside the TyCon (via tyConNullaryTy. But we have to build the TyConApp tc [] in that TyCon field; that's what mkNakedTyConTy is for.

mkTyVarTy :: TyVar -> Type

mkTyVarTys :: [TyVar] -> [Type]

mkTyCoVarTy :: TyCoVar -> Type

mkTyCoVarTys :: [TyCoVar] -> [Type]

mkFunTy :: HasDebugCallStack => FunTyFlag -> Mult -> Type -> Type -> Type infixr 3

mkNakedFunTy :: FunTyFlag -> Kind -> Kind -> Kind

mkVisFunTy :: HasDebugCallStack => Mult -> Type -> Type -> Type

mkScaledFunTys :: HasDebugCallStack => [Scaled Type] -> Type -> Type

mkInvisFunTy :: HasDebugCallStack => Type -> Type -> Type infixr 3

mkInvisFunTys :: HasDebugCallStack => [Type] -> Type -> Type

tcMkVisFunTy :: Mult -> Type -> Type -> Type

tcMkInvisFunTy :: TypeOrConstraint -> Type -> Type -> Type

tcMkScaledFunTys :: [Scaled Type] -> Type -> Type

mkForAllTy :: ForAllTyBinder -> Type -> Type

Like mkTyCoForAllTy, but does not check the occurrence of the binder See Note [Unused coercion variable in ForAllTy]

mkForAllTys :: [ForAllTyBinder] -> Type -> Type

Wraps foralls over the type using the provided TyCoVars from left to right

mkInvisForAllTys :: [InvisTVBinder] -> Type -> Type

Wraps foralls over the type using the provided InvisTVBinders from left to right

mkPiTy :: PiTyBinder -> Type -> Type

mkPiTys :: [PiTyBinder] -> Type -> Type

mkVisFunTyMany :: HasDebugCallStack => Type -> Type -> Type infixr 3

Make nested arrow types | Special, common, case: Arrow type with mult Many

mkVisFunTysMany :: [Type] -> Type -> Type

nonDetCmpTyLit :: TyLit -> TyLit -> Ordering

cmpTyLit :: TyLit -> TyLit -> Ordering

Functions over coercions

pickLR :: LeftOrRight -> (a, a) -> a

Analyzing types

data TyCoFolder env a

Constructors

TyCoFolder

Fields

tcf_view :: Type -> Maybe Type

tcf_tyvar :: env -> TyVar -> a

tcf_covar :: env -> CoVar -> a

tcf_hole :: env -> CoercionHole -> a

What to do with coercion holes. See Note [Coercion holes] in GHC.Core.TyCo.Rep.

tcf_tycobinder :: env -> TyCoVar -> ForAllTyFlag -> env

The returned env is used in the extended scope

foldTyCo :: Monoid a => TyCoFolder env a -> env -> (Type -> a, [Type] -> a, Coercion -> a, [Coercion] -> a)

noView :: Type -> Maybe Type

A view function that looks through nothing.

Sizes

typeSize :: Type -> Int

coercionSize :: Coercion -> Int

provSize :: UnivCoProvenance -> Int

Multiplicities

data Scaled a

A shorthand for data with an attached Mult element (the multiplicity).

Constructors

Scaled !Mult a

scaledMult :: Scaled a -> Mult

scaledThing :: Scaled a -> a

mapScaledType :: (Type -> Type) -> Scaled Type -> Scaled Type

Apply a function to both the Mult and the Type in a 'Scaled Type'

type Mult = Type

Mult is a type alias for Type.

Mult must contain Type because multiplicity variables are mere type variables (of kind Multiplicity) in Haskell. So the simplest implementation is to make Mult be Type.

Multiplicities can be formed with: - One: GHC.Types.One (= oneDataCon) - Many: GHC.Types.Many (= manyDataCon) - Multiplication: GHC.Types.MultMul (= multMulTyCon)

So that Mult feels a bit more structured, we provide pattern synonyms and smart constructors for these.