GHC.Core.Type

Description

Main functions for manipulating types and type-related things

Main data types representing Types

Types are any, but at least one, of:

Boxed

Iff its representation is a pointer to an object on the GC'd heap. Operationally, heap objects can be entered as a means of evaluation.

Lifted

Iff it has bottom as an element: An instance of a lifted type might diverge when evaluated. GHC Haskell's unboxed types are unlifted. An unboxed, but lifted type is not very useful. (Example: A byte-represented type, where evaluating 0xff computes the 12345678th collatz number modulo 0xff.) Only lifted types may be unified with a type variable.

Algebraic

Iff it is a type with one or more constructors, whether declared with data or newtype. An algebraic type is one that can be deconstructed with a case expression. There are algebraic types that are not lifted types, like unlifted data types or unboxed tuples.

Data

Iff it is a type declared with data, or a boxed tuple. There are also unlifted data types.

Primitive

Iff it is a built-in type that can't be expressed in Haskell.

Currently, all primitive types are unlifted, but that's not necessarily the case: for example, Int could be primitive.

Some primitive types are unboxed, such as Int#, whereas some are boxed but unlifted (such as ByteArray#). The only primitive types that we classify as algebraic are the unboxed tuples.

Some examples of type classifications that may make this a bit clearer are:

Type          primitive       boxed           lifted          algebraic
-----------------------------------------------------------------------------
Int#          Yes             No              No              No
ByteArray#    Yes             Yes             No              No
(# a, b #)  Yes             No              No              Yes
(# a | b #) Yes             No              No              Yes
(  a, b  )    No              Yes             Yes             Yes
[a]           No              Yes             Yes             Yes

A source type is a type that is a separate type as far as the type checker is concerned, but which has a more low-level representation as far as Core-to-Core passes and the rest of the back end is concerned.

You don't normally have to worry about this, as the utility functions in this module will automatically convert a source into a representation type if they are spotted, to the best of its abilities. If you don't want this to happen, use the equivalent functions from the TcType module.

data Type

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, TyCoVarBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible Specificity

Required

Bundled Patterns

pattern Specified :: ArgFlag

pattern Inferred :: ArgFlag

data AnonArgFlag

The non-dependent version of ArgFlag. See Note [AnonArgFlag] Appears here partly so that it's together with its friends ArgFlag and ForallVisFlag, but also because it is used in IfaceType, rather early in the compilation chain

Constructors

VisArg
Used for (->): an ordinary non-dependent arrow. The argument is visible in source code.

InvisArg
Used for (=>): a non-dependent predicate arrow. The argument is invisible in source code.

data Specificity

Whether an Invisible argument may appear in source Haskell.

Constructors

InferredSpec
the argument may not appear in source Haskell, it is only inferred.

SpecifiedSpec
the argument may appear in source Haskell, but isn't required.

type KindOrType = Type

The key representation of types within the compiler

type PredType = Type

A type of the form p of constraint kind represents a value whose type is the Haskell predicate p, where a predicate is what occurs before the => in a Haskell type.

We use PredType as documentation to mark those types that we guarantee to have this kind.

It can be expanded into its representation, but:

The type checker must treat it as opaque

The rest of the compiler treats it as transparent

Consider these examples:

f :: (Eq a) => a -> Int
g :: (?x :: Int -> Int) => a -> Int
h :: (r\l) => {r} => {l::Int | r}

Here the Eq a and ?x :: Int -> Int and rl are all called "predicates"

type ThetaType = [PredType]

A collection of PredTypes

data Var

Variable

Essentially a typed Name, that may also contain some additional information about the Var and its use sites.

type TyVar = Var

Type or kind Variable

isTyVar :: Var -> Bool

Is this a type-level (i.e., computationally irrelevant, thus erasable) variable? Satisfies isTyVar = not . isId.

type TyCoVar = Id

Type or Coercion Variable

data TyCoBinder

A TyCoBinder represents an argument to a function. TyCoBinders can be dependent (Named) or nondependent (Anon). They may also be visible or not. See Note [TyCoBinders]

type TyCoVarBinder = VarBndr TyCoVar ArgFlag

Variable Binder

A TyCoVarBinder is the binder of a ForAllTy It's convenient to define this synonym here rather its natural home in GHC.Core.TyCo.Rep, because it's used in GHC.Core.DataCon.hs-boot

A TyVarBinder is a binder with only TyVar

type TyVarBinder = VarBndr TyVar ArgFlag

type Mult = Type

Mult is a type alias for Type.

Mult must contain Type because multiplicity variables are mere type variables (of kind Multiplicity) in Haskell. So the simplest implementation is to make Mult be Type.

Multiplicities can be formed with: - One: GHC.Types.One (= oneDataCon) - Many: GHC.Types.Many (= manyDataCon) - Multiplication: GHC.Types.MultMul (= multMulTyCon)

So that Mult feels a bit more structured, we provide pattern synonyms and smart constructors for these.

data Scaled a

A shorthand for data with an attached Mult element (the multiplicity).

type KnotTied ty = ty

A type labeled KnotTied might have knot-tied tycons in it. See Note [Type checking recursive type and class declarations] in GHC.Tc.TyCl

Constructing and deconstructing types

mkTyVarTy :: TyVar -> Type

mkTyVarTys :: [TyVar] -> [Type]

getTyVar :: String -> Type -> TyVar

Attempts to obtain the type variable underlying a Type, and panics with the given message if this is not a type variable type. See also getTyVar_maybe

getTyVar_maybe :: Type -> Maybe TyVar

Attempts to obtain the type variable underlying a Type

repGetTyVar_maybe :: Type -> Maybe TyVar

Attempts to obtain the type variable underlying a Type, without any expansion

getCastedTyVar_maybe :: Type -> Maybe (TyVar, CoercionN)

If the type is a tyvar, possibly under a cast, returns it, along with the coercion. Thus, the co is :: kind tv ~N kind ty

tyVarKind :: TyVar -> Kind

varType :: Var -> Kind

The type or kind of the Var in question

mkAppTy :: Type -> Type -> Type

Applies a type to another, as in e.g. k a

mkAppTys :: Type -> [Type] -> Type

splitAppTy :: Type -> (Type, Type)

Attempts to take a type application apart, as in splitAppTy_maybe, and panics if this is not possible

splitAppTys :: Type -> (Type, [Type])

Recursively splits a type as far as is possible, leaving a residual type being applied to and the type arguments applied to it. Never fails, even if that means returning an empty list of type applications.

repSplitAppTys :: HasDebugCallStack => Type -> (Type, [Type])

Like splitAppTys, but doesn't look through type synonyms

splitAppTy_maybe :: Type -> Maybe (Type, Type)

Attempt to take a type application apart, whether it is a function, type constructor, or plain type application. Note that type family applications are NEVER unsaturated by this!

repSplitAppTy_maybe :: HasDebugCallStack => Type -> Maybe (Type, Type)

Does the AppTy split as in splitAppTy_maybe, but assumes that any Core view stuff is already done

tcRepSplitAppTy_maybe :: Type -> Maybe (Type, Type)

Does the AppTy split as in tcSplitAppTy_maybe, but assumes that any coreView stuff is already done. Refuses to look through (c => t)

mkFunTy :: AnonArgFlag -> Mult -> Type -> Type -> Type infixr 3

mkVisFunTy :: Mult -> Type -> Type -> Type infixr 3

mkInvisFunTy :: Mult -> Type -> Type -> Type infixr 3

mkVisFunTys :: [Scaled Type] -> Type -> Type

Make nested arrow types

mkVisFunTyMany :: Type -> Type -> Type infixr 3

Special, common, case: Arrow type with mult Many

mkInvisFunTyMany :: Type -> Type -> Type infixr 3

mkVisFunTysMany :: [Type] -> Type -> Type

mkInvisFunTysMany :: [Type] -> Type -> Type

splitFunTy :: Type -> (Mult, Type, Type)

Attempts to extract the multiplicity, argument and result types from a type, and panics if that is not possible. See also splitFunTy_maybe

splitFunTy_maybe :: Type -> Maybe (Mult, Type, Type)

Attempts to extract the multiplicity, argument and result types from a type

splitFunTys :: Type -> ([Scaled Type], Type)

funResultTy :: Type -> Type

Extract the function result type and panic if that is not possible

funArgTy :: Type -> Type

Just like piResultTys but for a single argument Try not to iterate piResultTy, because it's inefficient to substitute one variable at a time; instead use 'piResultTys"

Extract the function argument type and panic if that is not possible

mkTyConApp :: TyCon -> [Type] -> Type

A key function: builds a TyConApp or FunTy as appropriate to its arguments. Applies its arguments to the constructor from left to right.

mkTyConTy :: TyCon -> Type

Create the plain type constructor type which has been applied to no type arguments at all.

tYPE :: Type -> Type

Given a RuntimeRep, applies TYPE to it. See Note [TYPE and RuntimeRep] in GHC.Builtin.Types.Prim.

tyConAppTyCon_maybe :: Type -> Maybe TyCon

The same as fst . splitTyConApp

tyConAppTyConPicky_maybe :: Type -> Maybe TyCon

Retrieve the tycon heading this type, if there is one. Does not look through synonyms.

tyConAppArgs_maybe :: Type -> Maybe [Type]

The same as snd . splitTyConApp

tyConAppTyCon :: Type -> TyCon

tyConAppArgs :: Type -> [Type]

splitTyConApp_maybe :: HasDebugCallStack => Type -> Maybe (TyCon, [Type])

Attempts to tease a type apart into a type constructor and the application of a number of arguments to that constructor

splitTyConApp :: Type -> (TyCon, [Type])

Attempts to tease a type apart into a type constructor and the application of a number of arguments to that constructor. Panics if that is not possible. See also splitTyConApp_maybe

tyConAppArgN :: Int -> Type -> Type

tcSplitTyConApp_maybe :: HasCallStack => Type -> Maybe (TyCon, [Type])

Split a type constructor application into its type constructor and applied types. Note that this may fail in the case of a FunTy with an argument of unknown kind FunTy (e.g. FunTy (a :: k) Int. since the kind of a isn't of the form TYPE rep). Consequently, you may need to zonk your type before using this function.

This does *not* split types headed with (=>), as that's not a TyCon in the type-checker.

If you only need the TyCon, consider using tcTyConAppTyCon_maybe.

splitListTyConApp_maybe :: Type -> Maybe Type

Attempts to tease a list type apart and gives the type of the elements if successful (looks through type synonyms)

repSplitTyConApp_maybe :: HasDebugCallStack => Type -> Maybe (TyCon, [Type])

Like splitTyConApp_maybe, but doesn't look through synonyms. This assumes the synonyms have already been dealt with.

tcRepSplitTyConApp_maybe :: HasDebugCallStack => Type -> Maybe (TyCon, [Type])

Like tcSplitTyConApp_maybe, but doesn't look through synonyms. This assumes the synonyms have already been dealt with.

Moreover, for a FunTy, it only succeeds if the argument types have enough info to extract the runtime-rep arguments that the funTyCon requires. This will usually be true; but may be temporarily false during canonicalization: see Note [Decomposing FunTy] in GHC.Tc.Solver.Canonical and Note [The Purely Kinded Type Invariant (PKTI)] in GHC.Tc.Gen.HsType, Wrinkle around FunTy

mkForAllTy :: TyCoVar -> ArgFlag -> Type -> Type

Like mkTyCoForAllTy, but does not check the occurrence of the binder See Note [Unused coercion variable in ForAllTy]

mkForAllTys :: [TyCoVarBinder] -> Type -> Type

Wraps foralls over the type using the provided TyCoVars from left to right

mkInvisForAllTys :: [InvisTVBinder] -> Type -> Type

Wraps foralls over the type using the provided InvisTVBinders from left to right

mkTyCoInvForAllTys :: [TyCoVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and Inferred, a common case

mkSpecForAllTy :: TyVar -> Type -> Type

Like mkForAllTy, but assumes the variable is dependent and Specified, a common case

mkSpecForAllTys :: [TyVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and Specified, a common case

mkVisForAllTys :: [TyVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and visible

mkTyCoInvForAllTy :: TyCoVar -> Type -> Type

Make a dependent forall over an Inferred variable

mkInfForAllTy :: TyVar -> Type -> Type

Like mkTyCoInvForAllTy, but tv should be a tyvar

mkInfForAllTys :: [TyVar] -> Type -> Type

Like mkTyCoInvForAllTys, but tvs should be a list of tyvar

splitForAllTyCoVars :: Type -> ([TyCoVar], Type)

Take a ForAllTy apart, returning the list of tycovars and the result type. This always succeeds, even if it returns only an empty list. Note that the result type returned may have free variables that were bound by a forall.

splitForAllReqTVBinders :: Type -> ([ReqTVBinder], Type)

Like splitForAllTyCoVars, but only splits ForAllTys with Required type variable binders. Furthermore, each returned tyvar is annotated with ().

splitForAllInvisTVBinders :: Type -> ([InvisTVBinder], Type)

Like splitForAllTyCoVars, but only splits ForAllTys with Invisible type variable binders. Furthermore, each returned tyvar is annotated with its Specificity.

splitForAllTyCoVarBinders :: Type -> ([TyCoVarBinder], Type)

Like splitPiTys but split off only named binders and returns TyCoVarBinders rather than TyCoBinders

splitForAllTyCoVar_maybe :: Type -> Maybe (TyCoVar, Type)

Attempts to take a forall type apart, but only if it's a proper forall, with a named binder

splitForAllTyCoVar :: Type -> (TyCoVar, Type)

Take a forall type apart, or panics if that is not possible.

splitForAllTyVar_maybe :: Type -> Maybe (TyCoVar, Type)

Like splitForAllTyCoVar_maybe, but only returns Just if it is a tyvar binder.

splitForAllCoVar_maybe :: Type -> Maybe (TyCoVar, Type)

Like splitForAllTyCoVar_maybe, but only returns Just if it is a covar binder.

splitPiTy_maybe :: Type -> Maybe (TyCoBinder, Type)

Attempts to take a forall type apart; works with proper foralls and functions

splitPiTy :: Type -> (TyCoBinder, Type)

Takes a forall type apart, or panics

splitPiTys :: Type -> ([TyCoBinder], Type)

Split off all TyCoBinders to a type, splitting both proper foralls and functions

mkTyConBindersPreferAnon

Arguments

:: [TyVar]
binders

-> TyCoVarSet
free variables of result

-> [TyConBinder]

Given a list of type-level vars and the free vars of a result kind, makes TyCoBinders, preferring anonymous binders if the variable is, in fact, not dependent. e.g. mkTyConBindersPreferAnon (k:*),(b:k),(c:k) We want (k:*) Named, (b:k) Anon, (c:k) Anon

All non-coercion binders are visible.

mkPiTy :: TyCoBinder -> Type -> Type

mkPiTys :: [TyCoBinder] -> Type -> Type

piResultTy :: HasDebugCallStack => Type -> Type -> Type

piResultTys :: HasDebugCallStack => Type -> [Type] -> Type

(piResultTys f_ty [ty1, .., tyn]) gives the type of (f ty1 .. tyn) where f :: f_ty piResultTys is interesting because: 1. f_ty may have more for-alls than there are args 2. Less obviously, it may have fewer for-alls For case 2. think of: piResultTys (forall a.a) [forall b.b, Int] This really can happen, but only (I think) in situations involving undefined. For example: undefined :: forall a. a Term: undefined (forall b. b->b) Int This term should have type (Int -> Int), but notice that there are more type args than foralls in undefineds type.

applyTysX :: [TyVar] -> Type -> [Type] -> Type

dropForAlls :: Type -> Type

Drops all ForAllTys

mkFamilyTyConApp :: TyCon -> [Type] -> Type

Given a family instance TyCon and its arg types, return the corresponding family type. E.g:

data family T a
data instance T (Maybe b) = MkT b

Where the instance tycon is :RTL, so:

mkFamilyTyConApp :RTL Int  =  T (Maybe Int)

buildSynTyCon

Arguments

:: Name

-> [KnotTied TyConBinder]

-> Kind
result kind

-> [Role]

-> KnotTied Type

-> TyCon

mkNumLitTy :: Integer -> Type

isNumLitTy :: Type -> Maybe Integer

Is this a numeric literal. We also look through type synonyms.

mkStrLitTy :: FastString -> Type

isStrLitTy :: Type -> Maybe FastString

Is this a symbol literal. We also look through type synonyms.

mkCharLitTy :: Char -> Type

isCharLitTy :: Type -> Maybe Char

Is this a char literal? We also look through type synonyms.

isLitTy :: Type -> Maybe TyLit

Is this a type literal (symbol, numeric, or char)?

isPredTy :: HasDebugCallStack => Type -> Bool

getRuntimeRep_maybe :: HasDebugCallStack => Type -> Maybe Type

Extract the RuntimeRep classifier of a type. For instance, getRuntimeRep_maybe Int = LiftedRep. Returns Nothing if this is not possible.

kindRep_maybe :: HasDebugCallStack => Kind -> Maybe Type

Given a kind (TYPE rr), extract its RuntimeRep classifier rr. For example, kindRep_maybe * = Just LiftedRep Returns Nothing if the kind is not of form (TYPE rr) Treats * and Constraint as the same

kindRep :: HasDebugCallStack => Kind -> Type

Extract the RuntimeRep classifier of a type from its kind. For example, kindRep * = LiftedRep; Panics if this is not possible. Treats * and Constraint as the same

mkCastTy :: Type -> Coercion -> Type

Make a CastTy. The Coercion must be nominal. Checks the Coercion for reflexivity, dropping it if it's reflexive. See Note [Respecting definitional equality] in GHC.Core.TyCo.Rep

mkCoercionTy :: Coercion -> Type

splitCastTy_maybe :: Type -> Maybe (Type, Coercion)

userTypeError_maybe :: Type -> Maybe Type

Is this type a custom user error? If so, give us the kind and the error message.

pprUserTypeErrorTy :: Type -> SDoc

Render a type corresponding to a user type error into a SDoc.

coAxNthLHS :: CoAxiom br -> Int -> Type

Get the type on the LHS of a coercion induced by a type/data family instance.

stripCoercionTy :: Type -> Coercion

splitInvisPiTys :: Type -> ([TyCoBinder], Type)

Like splitPiTys, but returns only *invisible* binders, including constraints. Stops at the first visible binder.

splitInvisPiTysN :: Int -> Type -> ([TyCoBinder], Type)

Same as splitInvisPiTys, but stop when - you have found n TyCoBinders, - or you run out of invisible binders

invisibleTyBndrCount :: Type -> Int

filterOutInvisibleTypes :: TyCon -> [Type] -> [Type]

Given a TyCon and a list of argument types, filter out any invisible (i.e., Inferred or Specified) arguments.

filterOutInferredTypes :: TyCon -> [Type] -> [Type]

Given a TyCon and a list of argument types, filter out any Inferred arguments.

partitionInvisibleTypes :: TyCon -> [Type] -> ([Type], [Type])

Given a TyCon and a list of argument types, partition the arguments into:

Inferred or Specified (i.e., invisible) arguments and

Required (i.e., visible) arguments

partitionInvisibles :: [(a, ArgFlag)] -> ([a], [a])

Given a list of things paired with their visibilities, partition the things into (invisible things, visible things).

tyConArgFlags :: TyCon -> [Type] -> [ArgFlag]

Given a TyCon and a list of argument types to which the TyCon is applied, determine each argument's visibility (Inferred, Specified, or Required).

Wrinkle: consider the following scenario:

T :: forall k. k -> k
tyConArgFlags T [forall m. m -> m -> m, S, R, Q]

After substituting, we get

T (forall m. m -> m -> m) :: (forall m. m -> m -> m) -> forall n. n -> n -> n

Thus, the first argument is invisible, S is visible, R is invisible again, and Q is visible.

appTyArgFlags :: Type -> [Type] -> [ArgFlag]

Given a Type and a list of argument types to which the Type is applied, determine each argument's visibility (Inferred, Specified, or Required).

Most of the time, the arguments will be Required, but not always. Consider f :: forall a. a -> Type. In f Type Bool, the first argument (Type) is Specified and the second argument (Bool) is Required. It is precisely this sort of higher-rank situation in which appTyArgFlags comes in handy, since f Type Bool would be represented in Core using AppTys. (See also #15792).

Analyzing types

data TyCoMapper env m

This describes how a "map" operation over a type/coercion should behave

Constructors

TyCoMapper

Fields

tcm_tyvar :: env -> TyVar -> m Type

tcm_covar :: env -> CoVar -> m Coercion

tcm_hole :: env -> CoercionHole -> m Coercion

What to do with coercion holes. See Note [Coercion holes] in GHC.Core.TyCo.Rep.

tcm_tycobinder :: env -> TyCoVar -> ArgFlag -> m (env, TyCoVar)

The returned env is used in the extended scope

tcm_tycon :: TyCon -> m TyCon

This is used only for TcTyCons a) To zonk TcTyCons b) To turn TcTyCons into TyCons. See Note [Type checking recursive type and class declarations] in GHC.Tc.TyCl

mapTyCo :: Monad m => TyCoMapper () m -> (Type -> m Type, [Type] -> m [Type], Coercion -> m Coercion, [Coercion] -> m [Coercion])

mapTyCoX :: Monad m => TyCoMapper env m -> (env -> Type -> m Type, env -> [Type] -> m [Type], env -> Coercion -> m Coercion, env -> [Coercion] -> m [Coercion])

data TyCoFolder env a

Constructors

TyCoFolder

Fields

tcf_view :: Type -> Maybe Type

tcf_tyvar :: env -> TyVar -> a

tcf_covar :: env -> CoVar -> a

tcf_hole :: env -> CoercionHole -> a

What to do with coercion holes. See Note [Coercion holes] in GHC.Core.TyCo.Rep.

tcf_tycobinder :: env -> TyCoVar -> ArgFlag -> env

The returned env is used in the extended scope

foldTyCo :: Monoid a => TyCoFolder env a -> env -> (Type -> a, [Type] -> a, Coercion -> a, [Coercion] -> a)

newTyConInstRhs :: TyCon -> [Type] -> Type

Unwrap one layer of newtype on a type constructor and its arguments, using an eta-reduced version of the newtype if possible. This requires tys to have at least newTyConInstArity tycon elements.

Binders

sameVis :: ArgFlag -> ArgFlag -> Bool

Do these denote the same level of visibility? Required arguments are visible, others are not. So this function equates Specified and Inferred. Used for printing.

mkTyCoVarBinder :: vis -> TyCoVar -> VarBndr TyCoVar vis

Make a named binder

mkTyCoVarBinders :: vis -> [TyCoVar] -> [VarBndr TyCoVar vis]

Make many named binders

mkTyVarBinder :: vis -> TyVar -> VarBndr TyVar vis

Make a named binder var should be a type variable

mkTyVarBinders :: vis -> [TyVar] -> [VarBndr TyVar vis]

Make many named binders Input vars should be type variables

tyVarSpecToBinders :: [VarBndr a Specificity] -> [VarBndr a ArgFlag]

mkAnonBinder :: AnonArgFlag -> Scaled Type -> TyCoBinder

Make an anonymous binder

isAnonTyCoBinder :: TyCoBinder -> Bool

Does this binder bind a variable that is not erased? Returns True for anonymous binders.

binderVar :: VarBndr tv argf -> tv

binderVars :: [VarBndr tv argf] -> [tv]

binderType :: VarBndr TyCoVar argf -> Type

binderArgFlag :: VarBndr tv argf -> argf

tyCoBinderType :: TyCoBinder -> Type

tyCoBinderVar_maybe :: TyCoBinder -> Maybe TyCoVar

tyBinderType :: TyBinder -> Type

binderRelevantType_maybe :: TyCoBinder -> Maybe Type

Extract a relevant type, if there is one.

isVisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is written in Haskell?

isInvisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is not written in Haskell?

isVisibleBinder :: TyCoBinder -> Bool

Does this binder bind a visible argument?

isInvisibleBinder :: TyCoBinder -> Bool

Does this binder bind an invisible argument?

isNamedBinder :: TyCoBinder -> Bool

tyConBindersTyCoBinders :: [TyConBinder] -> [TyCoBinder]

Common type constructors

funTyCon :: TyCon

The FUN type constructor.

FUN :: forall (m :: Multiplicity) ->
       forall {rep1 :: RuntimeRep} {rep2 :: RuntimeRep}.
       TYPE rep1 -> TYPE rep2 -> *

The runtime representations quantification is left inferred. This means they cannot be specified with -XTypeApplications.

This is a deliberate choice to allow future extensions to the function arrow. To allow visible application a type synonym can be defined:

type Arr :: forall (rep1 :: RuntimeRep) (rep2 :: RuntimeRep).
            TYPE rep1 -> TYPE rep2 -> Type
type Arr = FUN 'Many

unrestrictedFunTyCon :: TyCon

Predicates on types

isTyVarTy :: Type -> Bool

isFunTy :: Type -> Bool

Is this a function?

isCoercionTy :: Type -> Bool

isCoercionTy_maybe :: Type -> Maybe Coercion

isForAllTy :: Type -> Bool

Checks whether this is a proper forall (with a named binder)

isForAllTy_ty :: Type -> Bool

Like isForAllTy, but returns True only if it is a tyvar binder

isForAllTy_co :: Type -> Bool

Like isForAllTy, but returns True only if it is a covar binder

isPiTy :: Type -> Bool

Is this a function or forall?

isTauTy :: Type -> Bool

isFamFreeTy :: Type -> Bool

isCoVarType :: Type -> Bool

Does this type classify a core (unlifted) Coercion? At either role nominal or representational (t1 ~# t2) or (t1 ~R# t2) See Note [Types for coercions, predicates, and evidence] in GHC.Core.TyCo.Rep

isAtomicTy :: Type -> Bool

isValidJoinPointType :: JoinArity -> Type -> Bool

Determine whether a type could be the type of a join point of given total arity, according to the polymorphism rule. A join point cannot be polymorphic in its return type, since given join j a b x y z = e1 in e2, the types of e1 and e2 must be the same, and a and b are not in scope for e2. (See Note [The polymorphism rule of join points] in GHC.Core.) Returns False also if the type simply doesn't have enough arguments.

Note that we need to know how many arguments (type *and* value) the putative join point takes; for instance, if j :: forall a. a -> Int then j could be a binary join point returning an Int, but it could *not* be a unary join point returning a -> Int.

TODO: See Note [Excess polymorphism and join points]

tyConAppNeedsKindSig

Arguments

:: Bool
Should specified binders count towards injective positions in the kind of the TyCon? (If you're using visible kind applications, then you want True here.

-> TyCon

-> Int
The number of args the TyCon is applied to.

-> Bool
Does T t_1 ... t_n need a kind signature? (Where n is the number of arguments)

Does a TyCon (that is applied to some number of arguments) need to be ascribed with an explicit kind signature to resolve ambiguity if rendered as a source-syntax type? (See Note [When does a tycon application need an explicit kind signature?] for a full explanation of what this function checks for.)

Levity and boxity

isLiftedType_maybe :: HasDebugCallStack => Type -> Maybe Bool

Returns Just True if this type is surely lifted, Just False if it is surely unlifted, Nothing if we can't be sure (i.e., it is levity polymorphic), and panics if the kind does not have the shape TYPE r.

isLiftedTypeKind :: Kind -> Bool

This version considers Constraint to be the same as *. Returns True if the argument is equivalent to Type/Constraint and False otherwise. See Note [Kind Constraint and kind Type]

isUnliftedTypeKind :: Kind -> Bool

Returns True if the kind classifies unlifted types (like 'Int#') and False otherwise. Note that this returns False for levity-polymorphic kinds, which may be specialized to a kind that classifies unlifted types.

isBoxedTypeKind :: Kind -> Bool

Returns True if the kind classifies types which are allocated on the GC'd heap and False otherwise. Note that this returns False for levity-polymorphic kinds, which may be specialized to a kind that classifies AddrRep or even unboxed kinds.

pickyIsLiftedTypeKind :: Kind -> Bool

isLiftedRuntimeRep :: Type -> Bool

isUnliftedRuntimeRep :: Type -> Bool

isBoxedRuntimeRep :: Type -> Bool

See isBoxedRuntimeRep_maybe.

isLiftedLevity :: Type -> Bool

isUnliftedLevity :: Type -> Bool

isUnliftedType :: HasDebugCallStack => Type -> Bool

See Type for what an unlifted type is. Panics on levity polymorphic types; See mightBeUnliftedType for a more approximate predicate that behaves better in the presence of levity polymorphism.

isBoxedType :: Type -> Bool

See Type for what a boxed type is. Panics on levity polymorphic types; See mightBeUnliftedType for a more approximate predicate that behaves better in the presence of levity polymorphism.

mightBeUnliftedType :: Type -> Bool

Returns:

False if the type is guaranteed lifted or

True if it is unlifted, OR we aren't sure (e.g. in a levity-polymorphic case)

isUnboxedTupleType :: Type -> Bool

isUnboxedSumType :: Type -> Bool

isAlgType :: Type -> Bool

See Type for what an algebraic type is. Should only be applied to types, as opposed to e.g. partially saturated type constructors

isDataFamilyAppType :: Type -> Bool

Check whether a type is a data family type

isPrimitiveType :: Type -> Bool

Returns true of types that are opaque to Haskell.

isStrictType :: HasDebugCallStack => Type -> Bool

Computes whether an argument (or let right hand side) should be computed strictly or lazily, based only on its type. Currently, it's just isUnliftedType. Panics on levity-polymorphic types.

isLevityTy :: Type -> Bool

Is this the type Levity?

isLevityVar :: TyVar -> Bool

Is a tyvar of type Levity?

isRuntimeRepTy :: Type -> Bool

Is this the type RuntimeRep?

isRuntimeRepVar :: TyVar -> Bool

Is a tyvar of type RuntimeRep?

isRuntimeRepKindedTy :: Type -> Bool

Is this a type of kind RuntimeRep? (e.g. LiftedRep)

dropRuntimeRepArgs :: [Type] -> [Type]

Drops prefix of RuntimeRep constructors in TyConApps. Useful for e.g. dropping 'LiftedRep arguments of unboxed tuple TyCon applications:

dropRuntimeRepArgs [ 'LiftedRep, 'IntRep , String, Int# ] == [String, Int#]

getRuntimeRep :: HasDebugCallStack => Type -> Type

Extract the RuntimeRep classifier of a type. For instance, getRuntimeRep_maybe Int = LiftedRep. Panics if this is not possible.

Multiplicity

isMultiplicityTy :: Type -> Bool

Is this the type Multiplicity?

isMultiplicityVar :: TyVar -> Bool

Is a tyvar of type Multiplicity?

unrestricted :: a -> Scaled a

Scale a payload by Many

linear :: a -> Scaled a

Scale a payload by One

tymult :: a -> Scaled a

Scale a payload by Many; used for type arguments in core

mkScaled :: Mult -> a -> Scaled a

irrelevantMult :: Scaled a -> a

scaledSet :: Scaled a -> b -> Scaled b

pattern One :: Mult

pattern Many :: Mult

isOneDataConTy :: Mult -> Bool

isManyDataConTy :: Mult -> Bool

isLinearType :: Type -> Bool

isLinear t returns True of a if t is a type of (curried) function where at least one argument is linear (or otherwise non-unrestricted). We use this function to check whether it is safe to eta reduce an Id in CorePrep. It is always safe to return True, because True deactivates the optimisation.

Main data types representing Kinds

type Kind = Type

The key type representing kinds in the compiler.

Finding the kind of a type

typeKind :: HasDebugCallStack => Type -> Kind

tcTypeKind :: HasDebugCallStack => Type -> Kind

isTypeLevPoly :: Type -> Bool

Returns True if a type is levity polymorphic. Should be the same as (isKindLevPoly . typeKind) but much faster. Precondition: The type has kind (TYPE blah)

resultIsLevPoly :: Type -> Bool

Looking past all pi-types, is the end result potentially levity polymorphic? Example: True for (forall r (a :: TYPE r). String -> a) Example: False for (forall r1 r2 (a :: TYPE r1) (b :: TYPE r2). a -> b -> Type)

tcIsLiftedTypeKind :: Kind -> Bool

Is this kind equivalent to Type?

This considers Constraint to be distinct from Type. For a version that treats them as the same type, see isLiftedTypeKind.

tcIsConstraintKind :: Kind -> Bool

tcReturnsConstraintKind :: Kind -> Bool

tcIsBoxedTypeKind :: Kind -> Bool

Is this kind equivalent to TYPE (BoxedRep l) for some l :: Levity?

This considers Constraint to be distinct from Type. For a version that treats them as the same type, see isLiftedTypeKind.

tcIsRuntimeTypeKind :: Kind -> Bool

Is this kind equivalent to TYPE r (for some unknown r)?

This considers Constraint to be distinct from *.

Common Kind

liftedTypeKind :: Kind

unliftedTypeKind :: Kind

Type free variables

tyCoFVsOfType :: Type -> FV

The worker for tyCoFVsOfType and tyCoFVsOfTypeList. The previous implementation used unionVarSet which is O(n+m) and can make the function quadratic. It's exported, so that it can be composed with other functions that compute free variables. See Note [FV naming conventions] in GHC.Utils.FV.

Eta-expanded because that makes it run faster (apparently) See Note [FV eta expansion] in GHC.Utils.FV for explanation.

tyCoFVsBndr :: TyCoVarBinder -> FV -> FV

tyCoFVsVarBndr :: Var -> FV -> FV

tyCoFVsVarBndrs :: [Var] -> FV -> FV

tyCoVarsOfType :: Type -> TyCoVarSet

tyCoVarsOfTypes :: [Type] -> TyCoVarSet

tyCoVarsOfTypeDSet :: Type -> DTyCoVarSet

tyCoFVsOfType that returns free variables of a type in a deterministic set. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in GHC.Utils.FV.

coVarsOfType :: Type -> CoVarSet

coVarsOfTypes :: [Type] -> CoVarSet

anyFreeVarsOfType :: (TyCoVar -> Bool) -> Type -> Bool

anyFreeVarsOfTypes :: (TyCoVar -> Bool) -> [Type] -> Bool

noFreeVarsOfType :: Type -> Bool

splitVisVarsOfType :: Type -> Pair TyCoVarSet

Retrieve the free variables in this type, splitting them based on whether they are used visibly or invisibly. Invisible ones come first.

splitVisVarsOfTypes :: [Type] -> Pair TyCoVarSet

expandTypeSynonyms :: Type -> Type

Expand out all type synonyms. Actually, it'd suffice to expand out just the ones that discard type variables (e.g. type Funny a = Int) But we don't know which those are currently, so we just expand all.

expandTypeSynonyms only expands out type synonyms mentioned in the type, not in the kinds of any TyCon or TyVar mentioned in the type.

Keep this synchronized with synonymTyConsOfType

typeSize :: Type -> Int

occCheckExpand :: [Var] -> Type -> Maybe Type

Closing over kinds

closeOverKindsDSet :: DTyVarSet -> DTyVarSet

Add the kind variables free in the kinds of the tyvars in the given set. Returns a deterministic set.

closeOverKindsList :: [TyVar] -> [TyVar]

Add the kind variables free in the kinds of the tyvars in the given set. Returns a deterministically ordered list.

closeOverKinds :: TyCoVarSet -> TyCoVarSet

Well-scoped lists of variables

scopedSort :: [TyCoVar] -> [TyCoVar]

Do a topological sort on a list of tyvars, so that binders occur before occurrences E.g. given [ a::k, k::*, b::k ] it'll return a well-scoped list [ k::*, a::k, b::k ]

This is a deterministic sorting operation (that is, doesn't depend on Uniques).

It is also meant to be stable: that is, variables should not be reordered unnecessarily. This is specified in Note [ScopedSort] See also Note [Ordering of implicit variables] in GHC.Rename.HsType

tyCoVarsOfTypeWellScoped :: Type -> [TyVar]

Get the free vars of a type in scoped order

tyCoVarsOfTypesWellScoped :: [Type] -> [TyVar]

Get the free vars of types in scoped order

Type comparison

eqType :: Type -> Type -> Bool

Type equality on source types. Does not look through newtypes or PredTypes, but it does look through type synonyms. This first checks that the kinds of the types are equal and then checks whether the types are equal, ignoring casts and coercions. (The kind check is a recursive call, but since all kinds have type Type, there is no need to check the types of kinds.) See also Note [Non-trivial definitional equality] in GHC.Core.TyCo.Rep.

eqTypeX :: RnEnv2 -> Type -> Type -> Bool

Compare types with respect to a (presumably) non-empty RnEnv2.

eqTypes :: [Type] -> [Type] -> Bool

Type equality on lists of types, looking through type synonyms but not newtypes.

nonDetCmpType :: Type -> Type -> Ordering

nonDetCmpTypes :: [Type] -> [Type] -> Ordering

nonDetCmpTypeX :: RnEnv2 -> Type -> Type -> Ordering

nonDetCmpTypesX :: RnEnv2 -> [Type] -> [Type] -> Ordering

nonDetCmpTc :: TyCon -> TyCon -> Ordering

Compare two TyCons. NB: This should never see Constraint (as recognized by Kind.isConstraintKindCon) which is considered a synonym for Type in Core. See Note [Kind Constraint and kind Type] in GHC.Core.Type. See Note [nonDetCmpType nondeterminism]

eqVarBndrs :: RnEnv2 -> [Var] -> [Var] -> Maybe RnEnv2

Forcing evaluation of types

seqType :: Type -> ()

seqTypes :: [Type] -> ()

Other views onto Types

coreView :: Type -> Maybe Type

This function strips off the top layer only of a type synonym application (if any) its underlying representation type. Returns Nothing if there is nothing to look through. This function considers Constraint to be a synonym of Type.

By being non-recursive and inlined, this case analysis gets efficiently joined onto the case analysis that the caller is already doing

tcView :: Type -> Maybe Type

Gives the typechecker view of a type. This unwraps synonyms but leaves Constraint alone. c.f. coreView, which turns Constraint into Type. Returns Nothing if no unwrapping happens. See also Note [coreView vs tcView]

tyConsOfType :: Type -> UniqSet TyCon

All type constructors occurring in the type; looking through type synonyms, but not newtypes. When it finds a Class, it returns the class TyCon.

Main type substitution data types

type TvSubstEnv = TyVarEnv Type

A substitution of Types for TyVars and Kinds for KindVars

data TCvSubst

Type & coercion substitution

The following invariants must hold of a TCvSubst:

The in-scope set is needed only to guide the generation of fresh uniques

In particular, the kind of the type variables in the in-scope set is not relevant

The substitution is only applied ONCE! This is because in general such application will not reach a fixed point.

Constructors

TCvSubst InScopeSet TvSubstEnv CvSubstEnv

Manipulating type substitutions

emptyTvSubstEnv :: TvSubstEnv

emptyTCvSubst :: TCvSubst

mkEmptyTCvSubst :: InScopeSet -> TCvSubst

mkTCvSubst :: InScopeSet -> (TvSubstEnv, CvSubstEnv) -> TCvSubst

zipTvSubst :: HasDebugCallStack => [TyVar] -> [Type] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please!

mkTvSubstPrs :: [(TyVar, Type)] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please!

zipTCvSubst :: HasDebugCallStack => [TyCoVar] -> [Type] -> TCvSubst

notElemTCvSubst :: Var -> TCvSubst -> Bool

getTvSubstEnv :: TCvSubst -> TvSubstEnv

setTvSubstEnv :: TCvSubst -> TvSubstEnv -> TCvSubst

zapTCvSubst :: TCvSubst -> TCvSubst

getTCvInScope :: TCvSubst -> InScopeSet

getTCvSubstRangeFVs :: TCvSubst -> VarSet

Returns the free variables of the types in the range of a substitution as a non-deterministic set.

extendTCvInScope :: TCvSubst -> Var -> TCvSubst

extendTCvInScopeList :: TCvSubst -> [Var] -> TCvSubst

extendTCvInScopeSet :: TCvSubst -> VarSet -> TCvSubst

extendTCvSubst :: TCvSubst -> TyCoVar -> Type -> TCvSubst

extendCvSubst :: TCvSubst -> CoVar -> Coercion -> TCvSubst

extendTvSubst :: TCvSubst -> TyVar -> Type -> TCvSubst

extendTvSubstBinderAndInScope :: TCvSubst -> TyCoBinder -> Type -> TCvSubst

extendTvSubstList :: TCvSubst -> [Var] -> [Type] -> TCvSubst

extendTvSubstAndInScope :: TCvSubst -> TyVar -> Type -> TCvSubst

extendTCvSubstList :: TCvSubst -> [Var] -> [Type] -> TCvSubst

extendTvSubstWithClone :: TCvSubst -> TyVar -> TyVar -> TCvSubst

extendTCvSubstWithClone :: TCvSubst -> TyCoVar -> TyCoVar -> TCvSubst

isInScope :: Var -> TCvSubst -> Bool

composeTCvSubstEnv :: InScopeSet -> (TvSubstEnv, CvSubstEnv) -> (TvSubstEnv, CvSubstEnv) -> (TvSubstEnv, CvSubstEnv)

(compose env1 env2)(x) is env1(env2(x)); i.e. apply env2 then env1. It assumes that both are idempotent. Typically, env1 is the refinement to a base substitution env2

composeTCvSubst :: TCvSubst -> TCvSubst -> TCvSubst

Composes two substitutions, applying the second one provided first, like in function composition.

zipTyEnv :: HasDebugCallStack => [TyVar] -> [Type] -> TvSubstEnv

zipCoEnv :: HasDebugCallStack => [CoVar] -> [Coercion] -> CvSubstEnv

isEmptyTCvSubst :: TCvSubst -> Bool

unionTCvSubst :: TCvSubst -> TCvSubst -> TCvSubst

Performing substitution on types and kinds

substTy :: HasCallStack => TCvSubst -> Type -> Type

Substitute within a Type The substitution has to satisfy the invariants described in Note [The substitution invariant].

substTys :: HasCallStack => TCvSubst -> [Type] -> [Type]

Substitute within several Types The substitution has to satisfy the invariants described in Note [The substitution invariant].

substScaledTy :: HasCallStack => TCvSubst -> Scaled Type -> Scaled Type

substScaledTys :: HasCallStack => TCvSubst -> [Scaled Type] -> [Scaled Type]

substTyWith :: HasCallStack => [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst

substTysWith :: [TyVar] -> [Type] -> [Type] -> [Type]

Type substitution, see zipTvSubst

substTheta :: HasCallStack => TCvSubst -> ThetaType -> ThetaType

Substitute within a ThetaType The substitution has to satisfy the invariants described in Note [The substitution invariant].

substTyAddInScope :: TCvSubst -> Type -> Type

Substitute within a Type after adding the free variables of the type to the in-scope set. This is useful for the case when the free variables aren't already in the in-scope set or easily available. See also Note [The substitution invariant].

substTyUnchecked :: TCvSubst -> Type -> Type

Substitute within a Type disabling the sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substTysUnchecked :: TCvSubst -> [Type] -> [Type]

Substitute within several Types disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTysUnchecked to substTys and remove this function. Please don't use in new code.

substScaledTyUnchecked :: HasCallStack => TCvSubst -> Scaled Type -> Scaled Type

substScaledTysUnchecked :: TCvSubst -> [Scaled Type] -> [Scaled Type]

substThetaUnchecked :: TCvSubst -> ThetaType -> ThetaType

Substitute within a ThetaType disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substThetaUnchecked to substTheta and remove this function. Please don't use in new code.

substTyWithUnchecked :: [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substCoUnchecked :: TCvSubst -> Coercion -> Coercion

Substitute within a Coercion disabling sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substCoWithUnchecked :: [TyVar] -> [Type] -> Coercion -> Coercion

Coercion substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substTyVarBndr :: HasCallStack => TCvSubst -> TyVar -> (TCvSubst, TyVar)

substTyVarBndrs :: HasCallStack => TCvSubst -> [TyVar] -> (TCvSubst, [TyVar])

substTyVar :: TCvSubst -> TyVar -> Type

substTyVars :: TCvSubst -> [TyVar] -> [Type]

substVarBndr :: HasCallStack => TCvSubst -> TyCoVar -> (TCvSubst, TyCoVar)

substVarBndrs :: HasCallStack => TCvSubst -> [TyCoVar] -> (TCvSubst, [TyCoVar])

substTyCoBndr :: TCvSubst -> TyCoBinder -> (TCvSubst, TyCoBinder)

cloneTyVarBndr :: TCvSubst -> TyVar -> Unique -> (TCvSubst, TyVar)

cloneTyVarBndrs :: TCvSubst -> [TyVar] -> UniqSupply -> (TCvSubst, [TyVar])

lookupTyVar :: TCvSubst -> TyVar -> Maybe Type

Tidying type related things up for printing

tidyType :: TidyEnv -> Type -> Type

Tidy a Type

See Note [Strictness in tidyType and friends]

tidyTypes :: TidyEnv -> [Type] -> [Type]

Tidy a list of Types

See Note [Strictness in tidyType and friends]

tidyOpenType :: TidyEnv -> Type -> (TidyEnv, Type)

tidyOpenTypes :: TidyEnv -> [Type] -> (TidyEnv, [Type])

Grabs the free type variables, tidies them and then uses tidyType to work over the type itself

tidyOpenKind :: TidyEnv -> Kind -> (TidyEnv, Kind)

tidyVarBndr :: TidyEnv -> TyCoVar -> (TidyEnv, TyCoVar)

tidyVarBndrs :: TidyEnv -> [TyCoVar] -> (TidyEnv, [TyCoVar])

This tidies up a type for printing in an error message, or in an interface file.

It doesn't change the uniques at all, just the print names.

tidyFreeTyCoVars :: TidyEnv -> [TyCoVar] -> TidyEnv

Add the free TyVars to the env in tidy form, so that we can tidy the type they are free in

tidyOpenTyCoVar :: TidyEnv -> TyCoVar -> (TidyEnv, TyCoVar)

Treat a new TyCoVar as a binder, and give it a fresh tidy name using the environment if one has not already been allocated. See also tidyVarBndr

tidyOpenTyCoVars :: TidyEnv -> [TyCoVar] -> (TidyEnv, [TyCoVar])

tidyTyCoVarOcc :: TidyEnv -> TyCoVar -> TyCoVar

tidyTopType :: Type -> Type

Calls tidyType on a top-level type (i.e. with an empty tidying environment)

tidyKind :: TidyEnv -> Kind -> Kind

tidyTyCoVarBinder :: TidyEnv -> VarBndr TyCoVar vis -> (TidyEnv, VarBndr TyCoVar vis)

tidyTyCoVarBinders :: TidyEnv -> [VarBndr TyCoVar vis] -> (TidyEnv, [VarBndr TyCoVar vis])

Kinds

isConstraintKindCon :: TyCon -> Bool

Returns True for the TyCon of the Constraint kind.

classifiesTypeWithValues :: Kind -> Bool

Does this classify a type allowed to have values? Responds True to things like *, #, TYPE Lifted, TYPE v, Constraint.

True of any sub-kind of OpenTypeKind

isKindLevPoly :: Kind -> Bool

Tests whether the given kind (which should look like TYPE x) is something other than a constructor tree (that is, constructors at every node). E.g. True of TYPE k, TYPE (F Int) False of TYPE 'LiftedRep