GHC.Core.Unfold.Make

Description

Unfolding creation

Documentation

noUnfolding :: Unfolding

There is no known Unfolding

mkUnfolding :: UnfoldingOpts -> UnfoldingSource -> Bool -> Bool -> CoreExpr -> Unfolding

mkCoreUnfolding :: UnfoldingSource -> Bool -> CoreExpr -> UnfoldingGuidance -> Unfolding

mkFinalUnfolding :: UnfoldingOpts -> UnfoldingSource -> StrictSig -> CoreExpr -> Unfolding

mkSimpleUnfolding :: UnfoldingOpts -> CoreExpr -> Unfolding

mkWorkerUnfolding :: SimpleOpts -> (CoreExpr -> CoreExpr) -> Unfolding -> Unfolding

mkInlineUnfolding :: SimpleOpts -> CoreExpr -> Unfolding

Make an unfolding that may be used unsaturated (ug_unsat_ok = unSaturatedOk) and that is reported as having its manifest arity (the number of outer lambdas applications will resolve before doing any work).

mkInlineUnfoldingWithArity :: Arity -> SimpleOpts -> CoreExpr -> Unfolding

Make an unfolding that will be used once the RHS has been saturated to the given arity.

mkInlinableUnfolding :: SimpleOpts -> CoreExpr -> Unfolding

mkWwInlineRule :: SimpleOpts -> CoreExpr -> Arity -> Unfolding

mkCompulsoryUnfolding :: SimpleOpts -> CoreExpr -> Unfolding

Used for things that absolutely must be unfolded

mkCompulsoryUnfolding' :: CoreExpr -> Unfolding

Same as mkCompulsoryUnfolding but no simple optimiser pass is performed on the unfolding.

mkDFunUnfolding :: [Var] -> DataCon -> [CoreExpr] -> Unfolding

specUnfolding :: SimpleOpts -> [Var] -> (CoreExpr -> CoreExpr) -> [CoreArg] -> Unfolding -> Unfolding