GHC.CoreToStg.Prep

Documentation

data CorePrepConfig

Constructors

CorePrepConfig

Fields

cp_catchNonexhaustiveCases :: !Bool

Whether to generate a default alternative with `error` in these cases. This is helpful when debugging demand analysis or type checker bugs which can sometimes manifest as segmentation faults.

cp_convertNumLit :: !(LitNumType -> Integer -> Maybe CoreExpr)

Convert some numeric literals (Integer, Natural) into their final Core form.

data CorePrepPgmConfig

Constructors

CorePrepPgmConfig

Fields

cpPgm_endPassConfig :: !EndPassConfig

cpPgm_generateDebugInfo :: !Bool

corePrepPgm :: Logger -> CorePrepConfig -> CorePrepPgmConfig -> Module -> ModLocation -> CoreProgram -> [TyCon] -> IO CoreProgram

corePrepExpr :: Logger -> CorePrepConfig -> CoreExpr -> IO CoreExpr

mkConvertNumLiteral :: Platform -> HomeUnit -> (Name -> IO TyThing) -> IO (LitNumType -> Integer -> Maybe CoreExpr)

Create a function that converts Bignum literals into their final CoreExpr