GHC.Data.FastString

Description

There are two principal string types used internally by GHC:

FastString

A compact, hash-consed, representation of character strings.

Generated by fsLit.

You can get a Unique from them.

Equality test is O(1) (it uses the Unique).

Comparison is O(1) or O(n):

O(n) but deterministic with lexical comparison (lexicalCompareFS)

O(1) but non-deterministic with Unique comparison (uniqCompareFS)

Turn into SDoc with ftext.

PtrString

Pointer and size of a Latin-1 encoded string.

Practically no operations.

Outputting them is fast.

Generated by sLit.

Turn into SDoc with ptext

Requires manual memory management. Improper use may lead to memory leaks or dangling pointers.

It assumes Latin-1 as the encoding, therefore it cannot represent arbitrary Unicode strings.

Use PtrString unless you want the facilities of FastString.

ByteString

bytesFS :: FastString -> ByteString

Gives the Modified UTF-8 encoded bytes corresponding to a FastString

fastStringToByteString :: FastString -> ByteString

Deprecated: Use bytesFS instead

Gives the Modified UTF-8 encoded bytes corresponding to a FastString

mkFastStringByteString :: ByteString -> FastString

Create a FastString by copying an existing ByteString

fastZStringToByteString :: FastZString -> ByteString

unsafeMkByteString :: String -> ByteString

ShortByteString

fastStringToShortByteString :: FastString -> ShortByteString

mkFastStringShortByteString :: ShortByteString -> FastString

Create a FastString from an existing ShortByteString without copying.

FastZString

data FastZString

hPutFZS :: Handle -> FastZString -> IO ()

zString :: FastZString -> String

lengthFZS :: FastZString -> Int

FastStrings

data FastString

A FastString is a UTF-8 encoded string together with a unique ID. All FastStrings are stored in a global hashtable to support fast O(1) comparison.

It is also associated with a lazy reference to the Z-encoding of this string which is used by the compiler internally.

Constructors

FastString

Fields

uniq :: !Int

n_chars :: !Int

fs_sbs :: !ShortByteString

fs_zenc :: FastZString

Lazily computed Z-encoding of this string. See Note [Z-Encoding] in GHC.Utils.Encoding.

Since FastStrings are globally memoized this is computed at most once for any given string.

newtype NonDetFastString

Non-deterministic FastString

This is a simple FastString wrapper with an Ord instance using uniqCompareFS (i.e. which compares FastStrings on their Uniques). Hence it is not deterministic from one run to the other.

Constructors

NonDetFastString FastString

newtype LexicalFastString

Lexical FastString

This is a simple FastString wrapper with an Ord instance using lexicalCompareFS (i.e. which compares FastStrings on their String representation). Hence it is deterministic from one run to the other.

Constructors

LexicalFastString FastString

Construction

fsLit :: String -> FastString

mkFastString :: String -> FastString

Creates a UTF-8 encoded FastString from a String

mkFastStringBytes :: Ptr Word8 -> Int -> FastString

mkFastStringByteList :: [Word8] -> FastString

Creates a FastString from a UTF-8 encoded [Word8]

mkFastString# :: Addr# -> FastString

Deconstruction

unpackFS :: FastString -> String

Unpacks and decodes the FastString

unconsFS :: FastString -> Maybe (Char, FastString)

Encoding

zEncodeFS :: FastString -> FastZString

Returns a Z-encoded version of a FastString. This might be the original, if it was already Z-encoded. The first time this function is applied to a particular FastString, the results are memoized.

Operations

uniqueOfFS :: FastString -> Int

lengthFS :: FastString -> Int

Returns the length of the FastString in characters

nullFS :: FastString -> Bool

Returns True if the FastString is empty

appendFS :: FastString -> FastString -> FastString

headFS :: FastString -> Char

concatFS :: [FastString] -> FastString

consFS :: Char -> FastString -> FastString

nilFS :: FastString

isUnderscoreFS :: FastString -> Bool

lexicalCompareFS :: FastString -> FastString -> Ordering

Compare FastString lexically

If you don't care about the lexical ordering, use uniqCompareFS instead.

uniqCompareFS :: FastString -> FastString -> Ordering

Compare FastString by their Unique (not lexically).

Much cheaper than lexicalCompareFS but non-deterministic!

Outputting

hPutFS :: Handle -> FastString -> IO ()

Outputs a FastString with no decoding at all, that is, you get the actual bytes in the FastString written to the Handle.

Internal

getFastStringTable :: IO [[[FastString]]]

getFastStringZEncCounter :: IO Int

PtrStrings

data PtrString

A PtrString is a pointer to some array of Latin-1 encoded chars.

Constructors

PtrString !(Ptr Word8) !Int

Construction

sLit :: String -> PtrString

mkPtrString# :: Addr# -> PtrString

Wrap an unboxed address into a PtrString.

mkPtrString :: String -> PtrString

Encode a String into a newly allocated PtrString using Latin-1 encoding. The original string must not contain non-Latin-1 characters (above codepoint 0xff).

Deconstruction

unpackPtrString :: PtrString -> String

Decode a PtrString back into a String using Latin-1 encoding. This does not free the memory associated with PtrString.

Operations

lengthPS :: PtrString -> Int

Return the length of a PtrString