GHC.Data.FlatBag

Documentation

data FlatBag a

Store elements in a flattened representation.

A FlatBag is a data structure that stores an ordered list of elements in a flat structure, avoiding the overhead of a linked list. Use this data structure, if the code requires the following properties:

Elements are stored in a long-lived object, and benefit from a flattened representation.

The FlatBag will be traversed but not extended or filtered.

The number of elements should be known.

Sharing of the empty case improves memory behaviour.

A FlagBag aims to have as little overhead as possible to store its elements. To achieve that, it distinguishes between the empty case, singleton, tuple and general case. Thus, we only pay for the additional three words of an Array if we have at least three elements.

Constructors

EmptyFlatBag

UnitFlatBag !a

TupleFlatBag !a !a

emptyFlatBag :: FlatBag a

Create an empty FlatBag.

The empty FlatBag is shared over all instances.

unitFlatBag :: a -> FlatBag a

Create a singleton FlatBag.

sizeFlatBag :: FlatBag a -> Word

Calculate the size of

elemsFlatBag :: FlatBag a -> [a]

Get all elements that are stored in the FlatBag.

mappendFlatBag :: FlatBag a -> FlatBag a -> FlatBag a

Combine two FlatBags.

The new FlatBag contains all elements from both FlatBags.

If one of the FlatBags is empty, the old FlatBag is reused.

Construction

fromList :: Word -> [a] -> FlatBag a

Store the list in a flattened memory representation, avoiding the memory overhead of a linked list.

The size n needs to be smaller or equal to the length of the list. If it is smaller than the length of the list, overflowing elements are discarded. It is undefined behaviour to set n to be bigger than the length of the list.

fromSizedSeq :: SizedSeq a -> FlatBag a

Convert a SizedSeq into its flattened representation. A 'FlatBag a' is more memory efficient than '[a]', if no further modification is necessary.