GHC.Data.Graph.Color

Description

Graph Coloring. This is a generic graph coloring library, abstracted over the type of the node keys, nodes and colors.

Documentation

module GHC.Data.Graph.Base

module GHC.Data.Graph.Ops

module GHC.Data.Graph.Ppr

colorGraph

Arguments

:: forall k cls color. (Uniquable k, Uniquable cls, Uniquable color, Eq cls, Ord k, Outputable k, Outputable cls, Outputable color)

=> Bool
whether to do iterative coalescing

-> Int
how many times we've tried to color this graph so far.

-> UniqFM cls (UniqSet color)
map of (node class -> set of colors available for this class).

-> Triv k cls color
fn to decide whether a node is trivially colorable.

-> (Graph k cls color -> k)
fn to choose a node to potentially leave uncolored if nothing is trivially colorable.

-> Graph k cls color
the graph to color.

-> (Graph k cls color, UniqSet k, UniqFM k k)

Try to color a graph with this set of colors. Uses Chaitin's algorithm to color the graph. The graph is scanned for nodes which are deamed 'trivially colorable'. These nodes are pushed onto a stack and removed from the graph. Once this process is complete the graph can be colored by removing nodes from the stack (ie in reverse order) and assigning them colors different to their neighbors.