GHC.Data.Graph.Directed

Documentation

data Graph node

graphFromEdgedVerticesOrd :: Ord key => [Node key payload] -> Graph (Node key payload)

graphFromEdgedVerticesUniq :: Uniquable key => [Node key payload] -> Graph (Node key payload)

data SCC vertex

Strongly connected component.

Constructors

AcyclicSCC vertex
A single vertex that is not in any cycle.

CyclicSCC [vertex]
A maximal set of mutually reachable vertices.

data Node key payload

Representation for nodes of the Graph.

The payload is user data, just carried around in this module

The key is the node identifier. Key has an Ord instance for performance reasons.

The [key] are the dependencies of the node; it's ok to have extra keys in the dependencies that are not the key of any Node in the graph

Constructors

DigraphNode

Fields

node_payload :: payload

User data

node_key :: key

User defined node id

node_dependencies :: [key]

Dependencies/successors of the node

flattenSCC :: SCC vertex -> [vertex]

The vertices of a strongly connected component.

flattenSCCs :: [SCC a] -> [a]

The vertices of a list of strongly connected components.

stronglyConnCompG :: Graph node -> [SCC node]

topologicalSortG :: Graph node -> [node]

verticesG :: Graph node -> [node]

edgesG :: Graph node -> [Edge node]

hasVertexG :: Graph node -> node -> Bool

reachableG :: Graph node -> node -> [node]

reachablesG :: Graph node -> [node] -> [node]

Given a list of roots return all reachable nodes.

transposeG :: Graph node -> Graph node

emptyG :: Graph node -> Bool

findCycle :: forall payload key. Ord key => [Node key payload] -> Maybe [payload]

Find a reasonably short cycle a->b->c->a, in a strongly connected component. The input nodes are presumed to be a SCC, so you can start anywhere.

stronglyConnCompFromEdgedVerticesOrd :: Ord key => [Node key payload] -> [SCC payload]

stronglyConnCompFromEdgedVerticesOrdR :: Ord key => [Node key payload] -> [SCC (Node key payload)]

stronglyConnCompFromEdgedVerticesUniq :: Uniquable key => [Node key payload] -> [SCC payload]

stronglyConnCompFromEdgedVerticesUniqR :: Uniquable key => [Node key payload] -> [SCC (Node key payload)]

data EdgeType

Edge direction based on DFS Classification

Constructors

Forward

Cross

Backward
Loop back towards the root node. Eg backjumps in loops

SelfLoop
v -> v

classifyEdges :: forall key. Uniquable key => key -> (key -> [key]) -> [(key, key)] -> [((key, key), EdgeType)]

Given a start vertex, a way to get successors from a node and a list of (directed) edges classify the types of edges.