GHC.Data.List.SetOps

Description

Set-like operations on lists

Avoid using them as much as possible

Documentation

unionLists :: (HasDebugCallStack, Outputable a, Eq a) => [a] -> [a] -> [a]

Assumes that the arguments contain no duplicates

minusList :: Ord a => [a] -> [a] -> [a]

Calculate the set difference of two lists. This is O((m + n) log n), where we subtract a list of n elements from a list of m elements.

Extremely short cases are handled specially: When m or n is 0, this takes O(1) time. When m is 1, it takes O(n) time.

type Assoc a b = [(a, b)]

A finite mapping based on equality and association lists.

assoc :: Eq a => String -> Assoc a b -> a -> b

assocMaybe :: Eq a => Assoc a b -> a -> Maybe b

Lookup key, fail gracefully using Nothing if not found.

assocUsing :: (a -> a -> Bool) -> String -> Assoc a b -> a -> b

assocDefault :: Eq a => b -> Assoc a b -> a -> b

assocDefaultUsing :: (a -> a -> Bool) -> b -> Assoc a b -> a -> b

hasNoDups :: Eq a => [a] -> Bool

removeDups :: (a -> a -> Ordering) -> [a] -> ([a], [NonEmpty a])

findDupsEq :: (a -> a -> Bool) -> [a] -> [NonEmpty a]

equivClasses :: (a -> a -> Ordering) -> [a] -> [NonEmpty a]

getNth :: Outputable a => [a] -> Int -> a