GHC.Data.Maybe

Documentation

module Data.Maybe

data MaybeErr err val

Constructors

Succeeded val

Failed err

failME :: err -> MaybeErr err val

isSuccess :: MaybeErr err val -> Bool

orElse :: Maybe a -> a -> a infixr 4

Flipped version of fromMaybe, useful for chaining.

firstJust :: Maybe a -> Maybe a -> Maybe a

firstJusts :: Foldable f => f (Maybe a) -> Maybe a

Takes a list of Maybes and returns the first Just if there is one, or Nothing otherwise.

firstJustsM :: (Monad m, Foldable f) => f (m (Maybe a)) -> m (Maybe a)

Takes computations returnings Maybes; tries each one in order. The first one to return a Just wins. Returns Nothing if all computations return Nothing.

whenIsJust :: Monad m => Maybe a -> (a -> m ()) -> m ()

expectJust :: HasCallStack => String -> Maybe a -> a

rightToMaybe :: Either a b -> Maybe b

MaybeT

newtype MaybeT (m :: Type -> Type) a

The parameterizable maybe monad, obtained by composing an arbitrary monad with the Maybe monad.

Computations are actions that may produce a value or exit.

The return function yields a computation that produces that value, while >>= sequences two subcomputations, exiting if either computation does.

Constructors

MaybeT

Fields

runMaybeT :: m (Maybe a)

liftMaybeT :: Monad m => m a -> MaybeT m a

tryMaybeT :: IO a -> MaybeT IO a

Try performing an IO action, failing on error.