GHC.Data.OrdList

Description

Provide trees (of instructions), so that lists of instructions can be appended in linear time.

Documentation

data OrdList a

pattern NilOL :: OrdList a

pattern ConsOL :: a -> OrdList a -> OrdList a

pattern SnocOL :: OrdList a -> a -> OrdList a

nilOL :: OrdList a

isNilOL :: OrdList a -> Bool

unitOL :: a -> OrdList a

appOL :: OrdList a -> OrdList a -> OrdList a infixl 5

consOL :: a -> OrdList a -> OrdList a infixr 5

snocOL :: OrdList a -> a -> OrdList a infixl 5

concatOL :: [OrdList a] -> OrdList a

lastOL :: OrdList a -> a

headOL :: OrdList a -> a

mapOL :: (a -> b) -> OrdList a -> OrdList b

mapOL' :: (a -> b) -> OrdList a -> OrdList b

fromOL :: OrdList a -> [a]

toOL :: [a] -> OrdList a

foldrOL :: (a -> b -> b) -> b -> OrdList a -> b

foldlOL :: (b -> a -> b) -> b -> OrdList a -> b

Strict left fold.

reverseOL :: OrdList a -> OrdList a

fromOLReverse :: OrdList a -> [a]

strictlyEqOL :: Eq a => OrdList a -> OrdList a -> Bool

Compare not only the values but also the structure of two lists

strictlyOrdOL :: Ord a => OrdList a -> OrdList a -> Ordering

Compare not only the values but also the structure of two lists