GHC.Data.UnionFind

Documentation

newtype Point s a

A variable which can be unified; alternately, this can be thought of as an equivalence class with a distinguished representative.

Constructors

Point (STRef s (Link s a))

writePoint :: Point s a -> Link s a -> ST s ()

Mutable write to a Point

readPoint :: Point s a -> ST s (Link s a)

Read the current value of Point.

data Link s a

The internal data structure for a Point, which either records the representative element of an equivalence class, or a link to the Point that actually stores the representative type.

Constructors

Info !(STRef s Int) !(STRef s a)

Link !(Point s a)

fresh :: a -> ST s (Point s a)

Create a fresh equivalence class with one element.

repr :: Point s a -> ST s (Point s a)

Flatten any chains of links, returning a Point which points directly to the canonical representation.

find :: Point s a -> ST s a

Return the canonical element of an equivalence class Point.

union :: Point s a -> Point s a -> ST s ()

Unify two equivalence classes, so that they share a canonical element. Keeps the descriptor of point2.

equivalent :: Point s a -> Point s a -> ST s Bool

Test if two points are in the same equivalence class.