GHC.Driver.Backend.Internal

Description

In versions of GHC up through 9.2, a Backend was represented only by its name. This module is meant to aid clients written against the GHC API, versions 9.2 and older. The module provides an alternative way to name any back end found in GHC 9.2. /Code within the GHC source tree should not import this module./ (#20927).

Only back ends found in version 9.2 have names.

Name of a back end

data BackendName

Constructors

NCG
Names the native code generator backend.

LLVM
Names the LLVM backend.

ViaC
Names the Via-C backend.

JavaScript
Names the JS backend.

Interpreter
Names the ByteCode interpreter.

NoBackend
Names the `-fno-code` backend.