GHC.Driver.CmdLine

Description

Command-line parser

This is an abstract command-line parser used by DynFlags.

(c) The University of Glasgow 2005

Documentation

processArgs

Arguments

:: Monad m

=> [Flag m]
cmdline parser spec

-> [Located String]
args

-> (FilePath -> EwM m [Located String])
response file handler

-> m ([Located String], [Err], Warns)

parseResponseFile :: forall (m :: Type -> Type). MonadIO m => FilePath -> EwM m [Located String]

Parse a response file into arguments.

data OptKind (m :: Type -> Type)

Constructors

NoArg (EwM m ())

HasArg (String -> EwM m ())

SepArg (String -> EwM m ())

Prefix (String -> EwM m ())

OptPrefix (String -> EwM m ())

OptIntSuffix (Maybe Int -> EwM m ())

IntSuffix (Int -> EwM m ())

Word64Suffix (Word64 -> EwM m ())

FloatSuffix (Float -> EwM m ())

PassFlag (String -> EwM m ())

AnySuffix (String -> EwM m ())

data GhcFlagMode

GHC flag modes describing when a flag has an effect.

Constructors

OnlyGhc
The flag only affects the non-interactive GHC

OnlyGhci
The flag only affects the interactive GHC

AllModes
The flag affects multiple ghc modes

HiddenFlag
This flag should not be seen in cli completion

data Flag (m :: Type -> Type)

Constructors

Flag

Fields

flagName :: String

flagOptKind :: OptKind m

flagGhcMode :: GhcFlagMode

defFlag :: forall (m :: Type -> Type). String -> OptKind m -> Flag m

defGhcFlag :: forall (m :: Type -> Type). String -> OptKind m -> Flag m

defGhciFlag :: forall (m :: Type -> Type). String -> OptKind m -> Flag m

defHiddenFlag :: forall (m :: Type -> Type). String -> OptKind m -> Flag m

hoistFlag :: (forall a. m a -> n a) -> Flag m -> Flag n

errorsToGhcException :: [(String, String)] -> GhcException

newtype Err

A command-line error message

Constructors

Err

Fields

errMsg :: Located String

type Warn = Located DriverMessage

A command-line warning message and the reason it arose

This used to be own type, but now it's just MsgEnvelope DriverMessage.

warnsToMessages :: DiagOpts -> [Warn] -> Messages DriverMessage

data EwM (m :: Type -> Type) a

runEwM :: EwM m a -> m (Errs, Warns, a)

addErr :: forall (m :: Type -> Type). Monad m => String -> EwM m ()

addWarn :: forall (m :: Type -> Type). Monad m => String -> EwM m ()

addFlagWarn :: forall (m :: Type -> Type). Monad m => DriverMessage -> EwM m ()

getArg :: forall (m :: Type -> Type). Monad m => EwM m String

getCurLoc :: forall (m :: Type -> Type). Monad m => EwM m SrcSpan

liftEwM :: Monad m => m a -> EwM m a