GHC.Driver.CodeOutput

Documentation

codeOutput

Arguments

:: Logger

-> TmpFs

-> DynFlags

-> UnitState

-> Module

-> FilePath

-> ModLocation

-> (a -> ForeignStubs)

-> [(ForeignSrcLang, FilePath)]
additional files to be compiled with the C compiler

-> [UnitId]

-> Stream IO RawCmmGroup a

-> IO (FilePath, (Bool, Maybe FilePath), [(ForeignSrcLang, FilePath)], a)

outputForeignStubs :: Logger -> TmpFs -> DynFlags -> UnitState -> Module -> ModLocation -> ForeignStubs -> IO (Bool, Maybe FilePath)

profilingInitCode :: Platform -> Module -> CollectedCCs -> CStub

Generate code to initialise cost centres

ipInitCode :: DynFlags -> Module -> [InfoProvEnt] -> CStub

Generate code to initialise info pointer origin See note [Mapping Info Tables to Source Positions]