GHC.Driver.Config.Diagnostic

Description

Functions for initialising error message printing configuration from the GHC session flags.

Documentation

initDiagOpts :: DynFlags -> DiagOpts

Initialise the general configuration for printing diagnostic messages For example, this configuration controls things like whether warnings are treated like errors.

initPrintConfig :: DynFlags -> DiagnosticOpts GhcMessage

Initialise the configuration for printing specific diagnostic messages

initPsMessageOpts :: DynFlags -> DiagnosticOpts PsMessage

initDsMessageOpts :: DynFlags -> DiagnosticOpts DsMessage

initTcMessageOpts :: DynFlags -> DiagnosticOpts TcRnMessage

initDriverMessageOpts :: DynFlags -> DiagnosticOpts DriverMessage