GHC.Driver.Config.Finder

Documentation

data FinderOpts

Locations and information the finder cares about.

Should be taken from DynFlags via initFinderOpts.

Constructors

FinderOpts

Fields

finder_importPaths :: [OsPath]

Where are we allowed to look for Modules and Source files

finder_lookupHomeInterfaces :: Bool

When looking up a home module:

True: search interface files (e.g. in '-c' mode)

False: search source files (e.g. in '--make' mode)

finder_bypassHiFileCheck :: Bool

Don't check that an imported interface file actually exists if it can only be at one location. The interface will be reported as InstalledFound even if the file doesn't exist, so this is only useful in specific cases (e.g. to generate dependencies with `ghc -M`)

finder_ways :: Ways

finder_enableSuggestions :: Bool

If we encounter unknown modules, should we suggest modules that have a similar name.

finder_workingDirectory :: Maybe OsPath

finder_thisPackageName :: Maybe FastString

finder_hiddenModules :: Set ModuleName

finder_reexportedModules :: Map ModuleName ModuleName

finder_hieDir :: Maybe OsPath

finder_hieSuf :: OsString

finder_hiDir :: Maybe OsPath

finder_hiSuf :: OsString

finder_dynHiSuf :: OsString

finder_objectDir :: Maybe OsPath

finder_objectSuf :: OsString

finder_dynObjectSuf :: OsString

finder_stubDir :: Maybe OsPath

initFinderOpts :: DynFlags -> FinderOpts

Create a new FinderOpts from DynFlags.