GHC.Driver.DynFlags

Dynamic flags and associated configuration types

data DumpFlag

Debugging flags

Constructors

Opt_D_dump_cmm

Opt_D_dump_cmm_from_stg

Opt_D_dump_cmm_raw

Opt_D_dump_cmm_verbose_by_proc

Opt_D_dump_cmm_verbose

Opt_D_dump_cmm_cfg

Opt_D_dump_cmm_cbe

Opt_D_dump_cmm_switch

Opt_D_dump_cmm_proc

Opt_D_dump_cmm_sp

Opt_D_dump_cmm_sink

Opt_D_dump_cmm_caf

Opt_D_dump_cmm_procmap

Opt_D_dump_cmm_split

Opt_D_dump_cmm_info

Opt_D_dump_cmm_cps

Opt_D_dump_cmm_thread_sanitizer

Opt_D_dump_cfg_weights
Dump the cfg used for block layout.

Opt_D_dump_asm

Opt_D_dump_asm_native

Opt_D_dump_asm_liveness

Opt_D_dump_asm_regalloc

Opt_D_dump_asm_regalloc_stages

Opt_D_dump_asm_conflicts

Opt_D_dump_asm_stats

Opt_D_dump_c_backend

Opt_D_dump_llvm

Opt_D_dump_js

Opt_D_dump_core_stats

Opt_D_dump_deriv

Opt_D_dump_ds

Opt_D_dump_ds_preopt

Opt_D_dump_foreign

Opt_D_dump_inlinings

Opt_D_dump_verbose_inlinings

Opt_D_dump_rule_firings

Opt_D_dump_rule_rewrites

Opt_D_dump_simpl_trace

Opt_D_dump_occur_anal

Opt_D_dump_parsed

Opt_D_dump_parsed_ast

Opt_D_dump_rn

Opt_D_dump_rn_ast

Opt_D_dump_simpl

Opt_D_dump_simpl_iterations

Opt_D_dump_spec

Opt_D_dump_spec_constr

Opt_D_dump_prep

Opt_D_dump_late_cc

Opt_D_dump_stg_from_core
Initial STG (CoreToStg output)

Opt_D_dump_stg_unarised
STG after unarise

Opt_D_dump_stg_cg
STG (after stg2stg)

Opt_D_dump_stg_tags
Result of tag inference analysis.

Opt_D_dump_stg_final
Final STG (before cmm gen)

Opt_D_dump_call_arity

Opt_D_dump_exitify

Opt_D_dump_dmdanal

Opt_D_dump_dmd_signatures

Opt_D_dump_cpranal

Opt_D_dump_cpr_signatures

Opt_D_dump_tc

Opt_D_dump_tc_ast

Opt_D_dump_hie

Opt_D_dump_types

Opt_D_dump_rules

Opt_D_dump_cse

Opt_D_dump_float_out

Opt_D_dump_float_in

Opt_D_dump_liberate_case

Opt_D_dump_static_argument_transformation

Opt_D_dump_worker_wrapper

Opt_D_dump_rn_trace

Opt_D_dump_rn_stats

Opt_D_dump_opt_cmm

Opt_D_dump_simpl_stats

Opt_D_dump_cs_trace

Opt_D_dump_tc_trace

Opt_D_dump_ec_trace

Opt_D_dump_if_trace

Opt_D_dump_splices

Opt_D_th_dec_file

Opt_D_dump_BCOs

Opt_D_dump_ticked

Opt_D_dump_rtti

Opt_D_source_stats

Opt_D_verbose_stg2stg

Opt_D_dump_hi

Opt_D_dump_hi_diffs

Opt_D_dump_mod_cycles

Opt_D_dump_mod_map

Opt_D_dump_timings

Opt_D_dump_view_pattern_commoning

Opt_D_verbose_core2core

Opt_D_dump_debug

Opt_D_dump_json

Opt_D_ppr_debug

Opt_D_no_debug_output

Opt_D_dump_faststrings

Opt_D_faststring_stats

Opt_D_ipe_stats

data GeneralFlag

Enumerates the simple on-or-off dynamic flags

Constructors

Opt_DumpToFile
Append dump output to files instead of stdout.

Opt_DumpWithWays
Use foo.ways.dumpFlag instead of foo.dumpFlag

Opt_D_dump_minimal_imports

Opt_DoCoreLinting

Opt_DoLinearCoreLinting

Opt_DoStgLinting

Opt_DoCmmLinting

Opt_DoAsmLinting

Opt_DoAnnotationLinting

Opt_DoBoundsChecking

Opt_NoLlvmMangler

Opt_FastLlvm

Opt_NoTypeableBinds

Opt_DistinctConstructorTables

Opt_InfoTableMap

Opt_InfoTableMapWithFallback

Opt_InfoTableMapWithStack

Opt_WarnIsError

Opt_ShowWarnGroups

Opt_HideSourcePaths

Opt_PrintExplicitForalls

Opt_PrintExplicitKinds

Opt_PrintExplicitCoercions

Opt_PrintExplicitRuntimeReps

Opt_PrintEqualityRelations

Opt_PrintAxiomIncomps

Opt_PrintUnicodeSyntax

Opt_PrintExpandedSynonyms

Opt_PrintPotentialInstances

Opt_PrintRedundantPromotionTicks

Opt_PrintTypecheckerElaboration

Opt_CallArity

Opt_Exitification

Opt_Strictness

Opt_LateDmdAnal

Opt_KillAbsence

Opt_KillOneShot

Opt_FullLaziness

Opt_FloatIn

Opt_LocalFloatOut
Enable floating out of let-bindings in the simplifier

Opt_LocalFloatOutTopLevel
Enable floating out of let-bindings at the top level in the simplifier N.B. See Note [RHS Floating]

Opt_LateSpecialise

Opt_Specialise

Opt_SpecialiseAggressively

Opt_CrossModuleSpecialise

Opt_PolymorphicSpecialisation

Opt_InlineGenerics

Opt_InlineGenericsAggressively

Opt_StaticArgumentTransformation

Opt_CSE

Opt_StgCSE

Opt_StgLiftLams

Opt_LiberateCase

Opt_SpecConstr

Opt_SpecConstrKeen

Opt_SpecialiseIncoherents

Opt_DoLambdaEtaExpansion

Opt_DoCleverArgEtaExpansion

Opt_IgnoreAsserts

Opt_DoEtaReduction

Opt_CaseMerge

Opt_CaseFolding

Opt_UnboxStrictFields

Opt_UnboxSmallStrictFields

Opt_DictsCheap

Opt_EnableRewriteRules

Opt_EnableThSpliceWarnings

Opt_RegsGraph

Opt_RegsIterative

Opt_PedanticBottoms

Opt_LlvmFillUndefWithGarbage

Opt_IrrefutableTuples

Opt_CmmSink

Opt_CmmStaticPred

Opt_CmmElimCommonBlocks

Opt_CmmControlFlow

Opt_AsmShortcutting

Opt_OmitYields

Opt_FunToThunk

Opt_DictsStrict

Opt_DmdTxDictSel
deprecated, no effect and behaviour is now default. Allowed switching of a special demand transformer for dictionary selectors

Opt_Loopification

Opt_CfgBlocklayout
Use the cfg based block layout algorithm.

Opt_WeightlessBlocklayout
Layout based on last instruction per block.

Opt_CprAnal

Opt_WorkerWrapper

Opt_WorkerWrapperUnlift
Do W/W split for unlifting even if we won't unbox anything.

Opt_SolveConstantDicts

Opt_AlignmentSanitisation

Opt_CatchNonexhaustiveCases

Opt_NumConstantFolding

Opt_CoreConstantFolding

Opt_FastPAPCalls

Opt_SpecEval

Opt_SpecEvalDictFun

Opt_DoTagInferenceChecks

Opt_SimplPreInlining

Opt_IgnoreInterfacePragmas

Opt_OmitInterfacePragmas

Opt_ExposeAllUnfoldings

Opt_ExposeOverloadedUnfoldings

Opt_KeepAutoRules
Keep auto-generated rules even if they seem to have become useless

Opt_WriteInterface

Opt_WriteHie

Opt_DisableJsMinifier
render JavaScript pretty-printed instead of minified (compacted)

Opt_DisableJsCsources
don't link C sources (compiled to JS) with Haskell code (compiled to JS)

Opt_AutoSccsOnIndividualCafs

Opt_ProfCountEntries

Opt_ProfLateInlineCcs

Opt_ProfLateCcs

Opt_ProfLateOverloadedCcs

Opt_ProfLateoverloadedCallsCCs

Opt_ProfManualCcs
Ignore manual SCC annotations

Opt_Pp

Opt_ForceRecomp

Opt_IgnoreOptimChanges

Opt_IgnoreHpcChanges

Opt_ExcessPrecision

Opt_EagerBlackHoling

Opt_OrigThunkInfo

Opt_NoHsMain

Opt_SplitSections

Opt_StgStats

Opt_HideAllPackages

Opt_HideAllPluginPackages

Opt_PrintBindResult

Opt_Haddock

Opt_HaddockOptions

Opt_BreakOnException

Opt_BreakOnError

Opt_PrintEvldWithShow

Opt_PrintBindContents

Opt_GenManifest

Opt_EmbedManifest

Opt_SharedImplib

Opt_BuildingCabalPackage

Opt_IgnoreDotGhci

Opt_GhciSandbox

Opt_InsertBreakpoints

Opt_GhciHistory

Opt_GhciLeakCheck

Opt_ValidateHie

Opt_LocalGhciHistory

Opt_NoIt

Opt_HelpfulErrors

Opt_DeferTypeErrors

Opt_DeferTypedHoles

Opt_DeferOutOfScopeVariables

Opt_PIC
-fPIC

Opt_PIE
-fPIE

Opt_PICExecutable
-pie

Opt_ExternalDynamicRefs

Opt_Ticky

Opt_Ticky_Allocd

Opt_Ticky_LNE

Opt_Ticky_Dyn_Thunk

Opt_Ticky_Tag

Opt_Ticky_AP
Use regular thunks even when we could use std ap thunks in order to get entry counts

Opt_CmmThreadSanitizer

Opt_RPath

Opt_RelativeDynlibPaths

Opt_CompactUnwind
-fcompact-unwind

Opt_Hpc

Opt_FamAppCache

Opt_ExternalInterpreter

Opt_OptimalApplicativeDo

Opt_VersionMacros

Opt_WholeArchiveHsLibs

Opt_SingleLibFolder

Opt_ExposeInternalSymbols

Opt_KeepCAFs

Opt_KeepGoing

Opt_ByteCode

Opt_ByteCodeAndObjectCode

Opt_UnoptimizedCoreForInterpreter

Opt_LinkRts

Opt_ErrorSpans

Opt_DeferDiagnostics

Opt_DiagnosticsAsJSON
Dump diagnostics as JSON

Opt_DiagnosticsShowCaret

Opt_PprCaseAsLet

Opt_PprShowTicks

Opt_ShowHoleConstraints

Opt_ShowValidHoleFits

Opt_SortValidHoleFits

Opt_SortBySizeHoleFits

Opt_SortBySubsumHoleFits

Opt_AbstractRefHoleFits

Opt_UnclutterValidHoleFits

Opt_ShowTypeAppOfHoleFits

Opt_ShowTypeAppVarsOfHoleFits

Opt_ShowDocsOfHoleFits

Opt_ShowTypeOfHoleFits

Opt_ShowProvOfHoleFits

Opt_ShowMatchesOfHoleFits

Opt_ShowLoadedModules

Opt_HexWordLiterals

Opt_SuppressCoercions

Opt_SuppressCoercionTypes

Opt_SuppressVarKinds

Opt_SuppressModulePrefixes

Opt_SuppressTypeApplications

Opt_SuppressIdInfo

Opt_SuppressUnfoldings

Opt_SuppressTypeSignatures

Opt_SuppressUniques

Opt_SuppressStgExts

Opt_SuppressStgReps

Opt_SuppressTicks

Opt_SuppressTimestamps
Suppress timestamps in dumps

Opt_SuppressCoreSizes
Suppress per binding Core size stats in dumps

Opt_ShowErrorContext

Opt_ObjectDeterminism

Opt_AutoLinkPackages

Opt_ImplicitImportQualified

Opt_KeepHscppFiles

Opt_KeepHiDiffs

Opt_KeepHcFiles

Opt_KeepSFiles

Opt_KeepTmpFiles

Opt_KeepRawTokenStream

Opt_KeepLlvmFiles

Opt_KeepHiFiles

Opt_KeepOFiles

Opt_BuildDynamicToo

Opt_WriteIfSimplifiedCore

Opt_UseBytecodeRatherThanObjects

Opt_DistrustAllPackages

Opt_PackageTrust

Opt_PluginTrustworthy

Opt_G_NoStateHack

Opt_G_NoOptCoercion

data WarningFlag

Constructors

Opt_WarnDuplicateExports

Opt_WarnDuplicateConstraints

Opt_WarnRedundantConstraints

Opt_WarnHiShadows

Opt_WarnImplicitPrelude

Opt_WarnIncompletePatterns

Opt_WarnIncompleteUniPatterns

Opt_WarnIncompletePatternsRecUpd

Opt_WarnOverflowedLiterals

Opt_WarnEmptyEnumerations

Opt_WarnMissingFields

Opt_WarnMissingImportList

Opt_WarnMissingMethods

Opt_WarnMissingSignatures

Opt_WarnMissingLocalSignatures

Opt_WarnNameShadowing

Opt_WarnOverlappingPatterns

Opt_WarnTypeDefaults

Opt_WarnMonomorphism

Opt_WarnUnusedTopBinds

Opt_WarnUnusedLocalBinds

Opt_WarnUnusedPatternBinds

Opt_WarnUnusedImports

Opt_WarnUnusedMatches

Opt_WarnUnusedTypePatterns

Opt_WarnUnusedForalls

Opt_WarnUnusedRecordWildcards

Opt_WarnRedundantBangPatterns

Opt_WarnRedundantRecordWildcards

Opt_WarnDeprecatedFlags

Opt_WarnMissingMonadFailInstances

Opt_WarnSemigroup

Opt_WarnDodgyExports

Opt_WarnDodgyImports

Opt_WarnOrphans

Opt_WarnAutoOrphans

Opt_WarnIdentities

Opt_WarnTabs

Opt_WarnUnrecognisedPragmas

Opt_WarnMisplacedPragmas

Opt_WarnDodgyForeignImports

Opt_WarnUnusedDoBind

Opt_WarnWrongDoBind

Opt_WarnAlternativeLayoutRuleTransitional

Opt_WarnUnsafe

Opt_WarnSafe

Opt_WarnTrustworthySafe

Opt_WarnMissedSpecs

Opt_WarnAllMissedSpecs

Opt_WarnUnsupportedCallingConventions

Opt_WarnUnsupportedLlvmVersion

Opt_WarnMissedExtraSharedLib

Opt_WarnInlineRuleShadowing

Opt_WarnTypedHoles

Opt_WarnPartialTypeSignatures

Opt_WarnMissingExportedSignatures

Opt_WarnUntickedPromotedConstructors

Opt_WarnDerivingTypeable

Opt_WarnDeferredTypeErrors

Opt_WarnDeferredOutOfScopeVariables

Opt_WarnNonCanonicalMonadInstances

Opt_WarnNonCanonicalMonadFailInstances

Opt_WarnNonCanonicalMonoidInstances

Opt_WarnMissingPatternSynonymSignatures

Opt_WarnUnrecognisedWarningFlags

Opt_WarnSimplifiableClassConstraints

Opt_WarnCPPUndef

Opt_WarnUnbangedStrictPatterns

Opt_WarnMissingHomeModules

Opt_WarnPartialFields

Opt_WarnMissingExportList

Opt_WarnInaccessibleCode

Opt_WarnStarIsType

Opt_WarnStarBinder

Opt_WarnImplicitKindVars

Opt_WarnSpaceAfterBang

Opt_WarnMissingDerivingStrategies

Opt_WarnPrepositiveQualifiedModule

Opt_WarnUnusedPackages

Opt_WarnInferredSafeImports

Opt_WarnMissingSafeHaskellMode

Opt_WarnCompatUnqualifiedImports

Opt_WarnDerivingDefaults

Opt_WarnInvalidHaddock

Opt_WarnOperatorWhitespaceExtConflict

Opt_WarnOperatorWhitespace

Opt_WarnAmbiguousFields

Opt_WarnImplicitLift

Opt_WarnMissingKindSignatures

Opt_WarnMissingPolyKindSignatures

Opt_WarnMissingExportedPatternSynonymSignatures

Opt_WarnRedundantStrictnessFlags

Opt_WarnForallIdentifier

Opt_WarnUnicodeBidirectionalFormatCharacters

Opt_WarnGADTMonoLocalBinds

Opt_WarnTypeEqualityOutOfScope

Opt_WarnTypeEqualityRequiresOperators

Opt_WarnLoopySuperclassSolve

Opt_WarnTermVariableCapture

Opt_WarnMissingRoleAnnotations

Opt_WarnImplicitRhsQuantification

Opt_WarnIncompleteExportWarnings

Opt_WarnIncompleteRecordSelectors

Opt_WarnBadlyStagedTypes

Opt_WarnInconsistentFlags

Opt_WarnDataKindsTC

Opt_WarnDeprecatedTypeAbstractions

Opt_WarnDefaultedExceptionContext

Opt_WarnViewPatternSignatures

data DiagnosticReason

The reason why a Diagnostic was emitted in the first place. Diagnostic messages are born within GHC with a very precise reason, which can be completely statically-computed (i.e. this is an error or a warning no matter what), or influenced by the specific state of the DynFlags at the moment of the creation of a new Diagnostic. For example, a parsing error is always going to be an error, whereas a 'WarningWithoutFlag Opt_WarnUnusedImports' might turn into an error due to '-Werror' or '-Werror=warn-unused-imports'. Interpreting a DiagnosticReason together with its associated Severity gives us the full picture.

Constructors

WarningWithoutFlag
Born as a warning.

WarningWithFlags !(NonEmpty WarningFlag)
Warning was enabled with the flag.

WarningWithCategory !WarningCategory
Warning was enabled with a custom category.

ErrorWithoutFlag
Born as an error.

Bundled Patterns

pattern WarningWithFlag :: WarningFlag -> DiagnosticReason
The single warning case DiagnosticReason is very common.

data Language

Constructors

Haskell98

Haskell2010

GHC2021

GHC2024

type FatalMessager = String -> IO ()

newtype FlushOut

Constructors

FlushOut (IO ())

data ProfAuto

What kind of {-# SCC #-} to add automatically

Constructors

NoProfAuto
no SCC annotations added

ProfAutoAll
top-level and nested functions are annotated

ProfAutoTop
top-level functions annotated only

ProfAutoExports
exported functions annotated only

ProfAutoCalls
annotate call-sites

hasPprDebug :: DynFlags -> Bool

hasNoDebugOutput :: DynFlags -> Bool

hasNoStateHack :: DynFlags -> Bool

hasNoOptCoercion :: DynFlags -> Bool

dopt :: DumpFlag -> DynFlags -> Bool

Test whether a DumpFlag is set

dopt_set :: DynFlags -> DumpFlag -> DynFlags

Set a DumpFlag

dopt_unset :: DynFlags -> DumpFlag -> DynFlags

Unset a DumpFlag

gopt :: GeneralFlag -> DynFlags -> Bool

Test whether a GeneralFlag is set

Note that dynamicNow (i.e., dynamic objects built with `-dynamic-too`) always implicitly enables Opt_PIC, Opt_ExternalDynamicRefs, and disables Opt_SplitSections.

gopt_set :: DynFlags -> GeneralFlag -> DynFlags

Set a GeneralFlag

gopt_unset :: DynFlags -> GeneralFlag -> DynFlags

Unset a GeneralFlag

wopt :: WarningFlag -> DynFlags -> Bool

Test whether a WarningFlag is set

wopt_set :: DynFlags -> WarningFlag -> DynFlags

Set a WarningFlag

wopt_unset :: DynFlags -> WarningFlag -> DynFlags

Unset a WarningFlag

wopt_fatal :: WarningFlag -> DynFlags -> Bool

Test whether a WarningFlag is set as fatal

wopt_set_fatal :: DynFlags -> WarningFlag -> DynFlags

Mark a WarningFlag as fatal (do not set the flag)

wopt_unset_fatal :: DynFlags -> WarningFlag -> DynFlags

Mark a WarningFlag as not fatal

wopt_set_all_custom :: DynFlags -> DynFlags

Enable all custom warning categories.

wopt_unset_all_custom :: DynFlags -> DynFlags

Disable all custom warning categories.

wopt_set_all_fatal_custom :: DynFlags -> DynFlags

Mark all custom warning categories as fatal (do not set the flags).

wopt_unset_all_fatal_custom :: DynFlags -> DynFlags

Mark all custom warning categories as non-fatal.

wopt_set_custom :: DynFlags -> WarningCategory -> DynFlags

Set a custom WarningCategory

wopt_unset_custom :: DynFlags -> WarningCategory -> DynFlags

Unset a custom WarningCategory

wopt_set_fatal_custom :: DynFlags -> WarningCategory -> DynFlags

Mark a custom WarningCategory as fatal (do not set the flag)

wopt_unset_fatal_custom :: DynFlags -> WarningCategory -> DynFlags

Mark a custom WarningCategory as not fatal

wopt_any_custom :: DynFlags -> Bool

Are there any custom warning categories enabled?

xopt :: Extension -> DynFlags -> Bool

Test whether a Extension is set

xopt_set :: DynFlags -> Extension -> DynFlags

Set a Extension

xopt_unset :: DynFlags -> Extension -> DynFlags

Unset a Extension

xopt_set_unlessExplSpec :: Extension -> (DynFlags -> Extension -> DynFlags) -> DynFlags -> DynFlags

Set or unset a Extension, unless it has been explicitly set or unset before.

xopt_DuplicateRecordFields :: DynFlags -> DuplicateRecordFields

xopt_FieldSelectors :: DynFlags -> FieldSelectors

lang_set :: DynFlags -> Maybe Language -> DynFlags

data DynamicTooState

Constructors

DT_Dont
Don't try to build dynamic objects too

DT_OK
Will still try to generate dynamic objects

DT_Dyn
Currently generating dynamic objects (in the backend)

dynamicTooState :: DynFlags -> DynamicTooState

setDynamicNow :: DynFlags -> DynFlags

data OnOff a

Constructors

On a

Off a

data DynFlags

Contains not only a collection of GeneralFlags but also a plethora of information relating to the compilation of a single file or GHC session

Constructors

DynFlags

Fields

ghcMode :: GhcMode

ghcLink :: GhcLink

backend :: !Backend

The backend to use (if any).

Whenever you change the backend, also make sure to set ghcLink to something sensible.

NoBackend can be used to avoid generating any output, however, note that:

If a program uses Template Haskell the typechecker may need to run code from an imported module. To facilitate this, code generation is enabled for modules imported by modules that use template haskell, using the default backend for the platform. See Note [-fno-code mode].

ghcNameVersion :: !GhcNameVersion

fileSettings :: !FileSettings

targetPlatform :: Platform

toolSettings :: !ToolSettings

platformMisc :: !PlatformMisc

rawSettings :: [(String, String)]

tmpDir :: TempDir

llvmOptLevel :: Int

LLVM optimisation level

verbosity :: Int

Verbosity level: see Note [Verbosity levels]

debugLevel :: Int

How much debug information to produce

simplPhases :: Int

Number of simplifier phases

maxSimplIterations :: Int

Max simplifier iterations

ruleCheck :: Maybe String

strictnessBefore :: [Int]

Additional demand analysis

parMakeCount :: Maybe ParMakeCount

The number of modules to compile in parallel If unspecified, compile with a single job.

enableTimeStats :: Bool

Enable RTS timing statistics?

ghcHeapSize :: Maybe Int

The heap size to set.

maxRelevantBinds :: Maybe Int

Maximum number of bindings from the type envt to show in type error messages

maxValidHoleFits :: Maybe Int

Maximum number of hole fits to show in typed hole error messages

maxRefHoleFits :: Maybe Int

Maximum number of refinement hole fits to show in typed hole error messages

refLevelHoleFits :: Maybe Int

Maximum level of refinement for refinement hole fits in typed hole error messages

maxUncoveredPatterns :: Int

Maximum number of unmatched patterns to show in non-exhaustiveness warnings

maxPmCheckModels :: Int

Soft limit on the number of models the pattern match checker checks a pattern against. A safe guard against exponential blow-up.

simplTickFactor :: Int

Multiplier for simplifier ticks

dmdUnboxWidth :: !Int

Whether DmdAnal should optimistically put an Unboxed demand on returned products with at most this number of fields

ifCompression :: Int

specConstrThreshold :: Maybe Int

Threshold for SpecConstr

specConstrCount :: Maybe Int

Max number of specialisations for any one function

specConstrRecursive :: Int

Max number of specialisations for recursive types Not optional; otherwise ForceSpecConstr can diverge.

binBlobThreshold :: Maybe Word

Binary literals (e.g. strings) whose size is above this threshold will be dumped in a binary file by the assembler code generator. 0 and Nothing disables this feature. See Config.

liberateCaseThreshold :: Maybe Int

Threshold for LiberateCase

floatLamArgs :: Maybe Int

Arg count for lambda floating See FloatOutSwitches

liftLamsRecArgs :: Maybe Int

Maximum number of arguments after lambda lifting a recursive function.

liftLamsNonRecArgs :: Maybe Int

Maximum number of arguments after lambda lifting a non-recursive function.

liftLamsKnown :: Bool

Lambda lift even when this turns a known call into an unknown call.

cmmProcAlignment :: Maybe Int

Align Cmm functions at this boundary or use default.

historySize :: Int

Simplification history size

importPaths :: [FilePath]

mainModuleNameIs :: ModuleName

mainFunIs :: Maybe String

reductionDepth :: IntWithInf

Typechecker maximum stack depth

solverIterations :: IntWithInf

Number of iterations in the constraints solver Typically only 1 is needed

givensFuel :: Int

Number of layers of superclass expansion for givens Should be < solverIterations See Note [Expanding Recursive Superclasses and ExpansionFuel]

wantedsFuel :: Int

Number of layers of superclass expansion for wanteds Should be < givensFuel See Note [Expanding Recursive Superclasses and ExpansionFuel]

qcsFuel :: Int

Number of layers of superclass expansion for quantified constraints Should be < givensFuel See Note [Expanding Recursive Superclasses and ExpansionFuel]

homeUnitId_ :: UnitId

Target home unit-id

homeUnitInstanceOf_ :: Maybe UnitId

Id of the unit to instantiate

homeUnitInstantiations_ :: [(ModuleName, Module)]

Module instantiations

workingDirectory :: Maybe FilePath

thisPackageName :: Maybe String

What the package is called, use with multiple home units

hiddenModules :: Set ModuleName

reexportedModules :: [ReexportedModule]

targetWays_ :: Ways

Target way flags from the command line

splitInfo :: Maybe (String, Int)

objectDir :: Maybe String

dylibInstallName :: Maybe String

hiDir :: Maybe String

hieDir :: Maybe String

stubDir :: Maybe String

dumpDir :: Maybe String

objectSuf_ :: String

hcSuf :: String

hiSuf_ :: String

hieSuf :: String

dynObjectSuf_ :: String

dynHiSuf_ :: String

outputFile_ :: Maybe String

dynOutputFile_ :: Maybe String

outputHi :: Maybe String

dynOutputHi :: Maybe String

dynLibLoader :: DynLibLoader

dynamicNow :: !Bool

Indicate if we are now generating dynamic output because of -dynamic-too. This predicate is used to query the appropriate fields (outputFile/dynOutputFile, ways, etc.)

dumpPrefix :: FilePath

This defaults to 'non-module'. It can be set by setDumpPrefix or 'ghc.GHCi.UI.runStmt' based on where its output is going.

dumpPrefixForce :: Maybe FilePath

Override the dumpPrefix set by setDumpPrefix or 'ghc.GHCi.UI.runStmt'. Set by -ddump-file-prefix

ldInputs :: [Option]

includePaths :: IncludeSpecs

libraryPaths :: [String]

frameworkPaths :: [String]

cmdlineFrameworks :: [String]

rtsOpts :: Maybe String

rtsOptsEnabled :: RtsOptsEnabled

rtsOptsSuggestions :: Bool

hpcDir :: String

Path to store the .mix files

pluginModNames :: [ModuleName]

the -fplugin flags given on the command line, in *reverse* order that they're specified on the command line.

pluginModNameOpts :: [(ModuleName, String)]

frontendPluginOpts :: [String]

the -ffrontend-opt flags given on the command line, in *reverse* order that they're specified on the command line.

externalPluginSpecs :: [ExternalPluginSpec]

External plugins loaded from shared libraries

depMakefile :: FilePath

depIncludePkgDeps :: Bool

depIncludeCppDeps :: Bool

depExcludeMods :: [ModuleName]

depSuffixes :: [String]

packageDBFlags :: [PackageDBFlag]

The -package-db flags given on the command line, In *reverse* order that they're specified on the command line. This is intended to be applied with the list of "initial" package databases derived from GHC_PACKAGE_PATH; see getUnitDbRefs.

ignorePackageFlags :: [IgnorePackageFlag]

The -ignore-package flags from the command line. In *reverse* order that they're specified on the command line.

packageFlags :: [PackageFlag]

The -package and -hide-package flags from the command-line. In *reverse* order that they're specified on the command line.

pluginPackageFlags :: [PackageFlag]

The -plugin-package-id flags from command line. In *reverse* order that they're specified on the command line.

trustFlags :: [TrustFlag]

The -trust and -distrust flags. In *reverse* order that they're specified on the command line.

packageEnv :: Maybe FilePath

Filepath to the package environment file (if overriding default)

dumpFlags :: EnumSet DumpFlag

generalFlags :: EnumSet GeneralFlag

warningFlags :: EnumSet WarningFlag

fatalWarningFlags :: EnumSet WarningFlag

customWarningCategories :: WarningCategorySet

fatalCustomWarningCategories :: WarningCategorySet

language :: Maybe Language

safeHaskell :: SafeHaskellMode

Safe Haskell mode

safeInfer :: Bool

safeInferred :: Bool

thOnLoc :: SrcSpan

newDerivOnLoc :: SrcSpan

deriveViaOnLoc :: SrcSpan

overlapInstLoc :: SrcSpan

incoherentOnLoc :: SrcSpan

pkgTrustOnLoc :: SrcSpan

warnSafeOnLoc :: SrcSpan

warnUnsafeOnLoc :: SrcSpan

trustworthyOnLoc :: SrcSpan

extensions :: [OnOff Extension]

extensionFlags :: EnumSet Extension

unfoldingOpts :: !UnfoldingOpts

Unfolding control See Note [Discounts and thresholds] in GHC.Core.Unfold

maxWorkerArgs :: Int

maxForcedSpecArgs :: Int

ghciHistSize :: Int

flushOut :: FlushOut

ghcVersionFile :: Maybe FilePath

haddockOptions :: Maybe String

ghciScripts :: [String]

GHCi scripts specified by -ghci-script, in reverse order

pprUserLength :: Int

pprCols :: Int

useUnicode :: Bool

useColor :: OverridingBool

canUseColor :: Bool

useErrorLinks :: OverridingBool

canUseErrorLinks :: Bool

colScheme :: Scheme

profAuto :: ProfAuto

what kind of {-# SCC #-} to add automatically

callerCcFilters :: [CallerCcFilter]

interactivePrint :: Maybe String

sseVersion :: Maybe SseVersion

Machine dependent flags (-m<blah> stuff)

bmiVersion :: Maybe BmiVersion

avx :: Bool

avx2 :: Bool

avx512cd :: Bool

avx512er :: Bool

avx512f :: Bool

avx512pf :: Bool

fma :: Bool

Enable FMA instructions.

maxInlineAllocSize :: Int

Max size, in bytes, of inline array allocations.

maxInlineMemcpyInsns :: Int

Only inline memcpy if it generates no more than this many pseudo (roughly: Cmm) instructions.

maxInlineMemsetInsns :: Int

Only inline memset if it generates no more than this many pseudo (roughly: Cmm) instructions.

reverseErrors :: Bool

Reverse the order of error messages in GHC/GHCi

maxErrors :: Maybe Int

Limit the maximum number of errors to show

initialUnique :: Word64

Unique supply configuration for testing build determinism

uniqueIncrement :: Int

cfgWeights :: Weights

Temporary: CFG Edge weights for fast iterations

data ParMakeCount

The type for the -jN argument, specifying that -j on its own represents using the number of machine processors.

Constructors

ParMakeThisMany Int
Use this many processors (-jn flag).

ParMakeNumProcessors
Use parallelism with as many processors as possible (-j flag without an argument).

ParMakeSemaphore FilePath
Use the specific semaphore sem to control parallelism (-jsem sem flag).

ways :: DynFlags -> Ways

class HasDynFlags (m :: Type -> Type) where

Methods

getDynFlags :: m DynFlags

class ContainsDynFlags t where

Methods

extractDynFlags :: t -> DynFlags

data RtsOptsEnabled

Constructors

RtsOptsNone

RtsOptsIgnore

RtsOptsIgnoreAll

RtsOptsSafeOnly

RtsOptsAll

data GhcMode

The GhcMode tells us whether we're doing multi-module compilation (controlled via the GHC API) or one-shot (single-module) compilation. This makes a difference primarily to the GHC.Unit.Finder: in one-shot mode we look for interface files for imported modules, but in multi-module mode we look for source files in order to check whether they need to be recompiled.

Constructors

CompManager
--make, GHCi, etc.

OneShot
ghc -c Foo.hs

MkDepend
ghc -M, see GHC.Unit.Finder for why we need this

isOneShot :: GhcMode -> Bool

data GhcLink

What to do in the link step, if there is one.

Constructors

NoLink
Don't link at all

LinkBinary
Link object code into a binary

LinkInMemory
Use the in-memory dynamic linker (works for both bytecode and object code).

LinkDynLib
Link objects into a dynamic lib (DLL on Windows, DSO on ELF platforms)

LinkStaticLib
Link objects into a static lib

LinkMergedObj
Link objects into a merged "GHCi object"

isNoLink :: GhcLink -> Bool

data PackageFlag

Flags for manipulating packages visibility.

Constructors

ExposePackage String PackageArg ModRenaming
-package, -package-id

HidePackage String
-hide-package

data PackageArg

We accept flags which make packages visible, but how they select the package varies; this data type reflects what selection criterion is used.

Constructors

PackageArg String
-package, by PackageName

UnitIdArg Unit
-package-id, by Unit

data ModRenaming

Represents the renaming that may be associated with an exposed package, e.g. the rns part of -package "foo (rns)".

Here are some example parsings of the package flags (where a string literal is punned to be a ModuleName:

-package foo is ModRenaming True []

-package foo () is ModRenaming False []

-package foo (A) is ModRenaming False [(A, A)]

-package foo (A as B) is ModRenaming False [(A, B)]

-package foo with (A as B) is ModRenaming True [(A, B)]

Constructors

ModRenaming

Fields

modRenamingWithImplicit :: Bool

Bring all exposed modules into scope?

modRenamings :: [(ModuleName, ModuleName)]

Bring module m into scope under name n.

packageFlagsChanged :: DynFlags -> DynFlags -> Bool

newtype IgnorePackageFlag

Flags for manipulating the set of non-broken packages.

Constructors

IgnorePackage String
-ignore-package

data TrustFlag

Flags for manipulating package trust.

Constructors

TrustPackage String
-trust

DistrustPackage String
-distrust

data PackageDBFlag

Constructors

PackageDB PkgDbRef

NoUserPackageDB

NoGlobalPackageDB

ClearPackageDBs

data PkgDbRef

Constructors

GlobalPkgDb

UserPkgDb

PkgDbPath FilePath

data Option

When invoking external tools as part of the compilation pipeline, we pass these a sequence of options on the command-line. Rather than just using a list of Strings, we use a type that allows us to distinguish between filepaths and 'other stuff'. The reason for this is that this type gives us a handle on transforming filenames, and filenames only, to whatever format they're expected to be on a particular platform.

Constructors

FileOption String String

Option String

showOpt :: Option -> String

data DynLibLoader

Constructors

Deployable

SystemDependent

positionIndependent :: DynFlags -> Bool

Are we building with -fPIE or -fPIC enabled?

optimisationFlags :: EnumSet GeneralFlag

The set of flags which affect optimisation for the purposes of recompilation avoidance. Specifically, these include flags which affect code generation but not the semantics of the program.

See Note [Ignoring some flag changes] in GHC.Iface.Recomp.Flags)

targetProfile :: DynFlags -> Profile

Get target profile

data ReexportedModule

Constructors

ReexportedModule

Fields

reexportFrom :: ModuleName

reexportTo :: ModuleName

Manipulating DynFlags

defaultDynFlags :: Settings -> DynFlags

The normal DynFlags. Note that they are not suitable for use in this form and must be fully initialized by runGhc first.

initDynFlags :: DynFlags -> IO DynFlags

Used by runGhc to partially initialize a new DynFlags value

defaultFatalMessager :: FatalMessager

defaultFlushOut :: FlushOut

optLevelFlags :: [([Int], GeneralFlag)]

languageExtensions :: Maybe Language -> [Extension]

The language extensions implied by the various language variants. When updating this be sure to update the flag documentation in docsusers_guideexts.

type TurnOnFlag = Bool

turnOn :: TurnOnFlag

turnOff :: TurnOnFlag

System tool settings and locations

programName :: DynFlags -> String

projectVersion :: DynFlags -> String

ghcUsagePath :: DynFlags -> FilePath

ghciUsagePath :: DynFlags -> FilePath

topDir :: DynFlags -> FilePath

toolDir :: DynFlags -> Maybe FilePath

versionedAppDir :: String -> ArchOS -> MaybeT IO FilePath

The directory for this version of ghc in the user's app directory The appdir used to be in ~/.ghc but to respect the XDG specification we want to move it under $XDG_DATA_HOME/ However, old tooling (like cabal) might still write package environments to the old directory, so we prefer that if a subdirectory of ~/.ghc with the correct target and GHC version suffix exists.

i.e. if ~.ghc$UNIQUE_SUBDIR exists we use that otherwise we use $XDG_DATA_HOME/$UNIQUE_SUBDIR

UNIQUE_SUBDIR is typically a combination of the target platform and GHC version

versionedFilePath :: ArchOS -> FilePath

extraGccViaCFlags :: DynFlags -> [String]

globalPackageDatabasePath :: DynFlags -> FilePath

Include specifications

data IncludeSpecs

Used to differentiate the scope an include needs to apply to. We have to split the include paths to avoid accidentally forcing recursive includes since -I overrides the system search paths. See #14312.

Constructors

IncludeSpecs

Fields

includePathsQuote :: [String]

includePathsGlobal :: [String]

includePathsQuoteImplicit :: [String]

See Note [Implicit include paths]

addGlobalInclude :: IncludeSpecs -> [String] -> IncludeSpecs

Append to the list of includes a path that shall be included using `-I` when the C compiler is called. These paths override system search paths.

addQuoteInclude :: IncludeSpecs -> [String] -> IncludeSpecs

Append to the list of includes a path that shall be included using `-iquote` when the C compiler is called. These paths only apply when quoted includes are used. e.g. #include "foo.h"

flattenIncludes :: IncludeSpecs -> [String]

Concatenate and flatten the list of global and quoted includes returning just a flat list of paths.

addImplicitQuoteInclude :: IncludeSpecs -> [String] -> IncludeSpecs

These includes are not considered while fingerprinting the flags for iface | See Note [Implicit include paths]

SDoc

initSDocContext :: DynFlags -> PprStyle -> SDocContext

Initialize the pretty-printing options

initDefaultSDocContext :: DynFlags -> SDocContext

Initialize the pretty-printing options using the default user style

initPromotionTickContext :: DynFlags -> PromotionTickContext

Platform features

isSse4_1Enabled :: DynFlags -> Bool

isSse4_2Enabled :: DynFlags -> Bool

isAvxEnabled :: DynFlags -> Bool

isAvx2Enabled :: DynFlags -> Bool

isAvx512cdEnabled :: DynFlags -> Bool

isAvx512erEnabled :: DynFlags -> Bool

isAvx512fEnabled :: DynFlags -> Bool

isAvx512pfEnabled :: DynFlags -> Bool

isFmaEnabled :: DynFlags -> Bool

isBmiEnabled :: DynFlags -> Bool

isBmi2Enabled :: DynFlags -> Bool