GHC.Driver.Errors

Documentation

warningsToMessages :: DynFlags -> WarningMessages -> (WarningMessages, ErrorMessages)

Converts a list of WarningMessages into a tuple where the second element contains only error, i.e. warnings that are considered fatal by GHC based on the input DynFlags.

printOrThrowWarnings :: Logger -> DynFlags -> Bag WarnMsg -> IO ()

Given a bag of warnings, turn them into an exception if -Werror is enabled, or print them out otherwise.

printBagOfErrors :: RenderableDiagnostic a => Logger -> DynFlags -> Bag (MsgEnvelope a) -> IO ()

isWarnMsgFatal :: DynFlags -> WarnMsg -> Maybe (Maybe WarningFlag)

Checks if given WarnMsg is a fatal warning.

handleFlagWarnings :: Logger -> DynFlags -> [Warn] -> IO ()