GHC.Driver.Flags

Documentation

data DumpFlag

Debugging flags

Constructors

Opt_D_dump_cmm

Opt_D_dump_cmm_from_stg

Opt_D_dump_cmm_raw

Opt_D_dump_cmm_verbose_by_proc

Opt_D_dump_cmm_verbose

Opt_D_dump_cmm_cfg

Opt_D_dump_cmm_cbe

Opt_D_dump_cmm_switch

Opt_D_dump_cmm_proc

Opt_D_dump_cmm_sp

Opt_D_dump_cmm_sink

Opt_D_dump_cmm_caf

Opt_D_dump_cmm_procmap

Opt_D_dump_cmm_split

Opt_D_dump_cmm_info

Opt_D_dump_cmm_cps

Opt_D_dump_cfg_weights
Dump the cfg used for block layout.

Opt_D_dump_asm

Opt_D_dump_asm_native

Opt_D_dump_asm_liveness

Opt_D_dump_asm_regalloc

Opt_D_dump_asm_regalloc_stages

Opt_D_dump_asm_conflicts

Opt_D_dump_asm_stats

Opt_D_dump_asm_expanded

Opt_D_dump_c_backend

Opt_D_dump_llvm

Opt_D_dump_core_stats

Opt_D_dump_deriv

Opt_D_dump_ds

Opt_D_dump_ds_preopt

Opt_D_dump_foreign

Opt_D_dump_inlinings

Opt_D_dump_rule_firings

Opt_D_dump_rule_rewrites

Opt_D_dump_simpl_trace

Opt_D_dump_occur_anal

Opt_D_dump_parsed

Opt_D_dump_parsed_ast

Opt_D_dump_rn

Opt_D_dump_rn_ast

Opt_D_dump_simpl

Opt_D_dump_simpl_iterations

Opt_D_dump_spec

Opt_D_dump_prep

Opt_D_dump_stg_from_core
Initial STG (CoreToStg output)

Opt_D_dump_stg_unarised
STG after unarise

Opt_D_dump_stg_final
Final STG (after stg2stg)

Opt_D_dump_call_arity

Opt_D_dump_exitify

Opt_D_dump_stranal

Opt_D_dump_str_signatures

Opt_D_dump_cpranal

Opt_D_dump_cpr_signatures

Opt_D_dump_tc

Opt_D_dump_tc_ast

Opt_D_dump_hie

Opt_D_dump_types

Opt_D_dump_rules

Opt_D_dump_cse

Opt_D_dump_worker_wrapper

Opt_D_dump_rn_trace

Opt_D_dump_rn_stats

Opt_D_dump_opt_cmm

Opt_D_dump_simpl_stats

Opt_D_dump_cs_trace

Opt_D_dump_tc_trace

Opt_D_dump_ec_trace

Opt_D_dump_if_trace

Opt_D_dump_vt_trace

Opt_D_dump_splices

Opt_D_th_dec_file

Opt_D_dump_BCOs

Opt_D_dump_ticked

Opt_D_dump_rtti

Opt_D_source_stats

Opt_D_verbose_stg2stg

Opt_D_dump_hi

Opt_D_dump_hi_diffs

Opt_D_dump_mod_cycles

Opt_D_dump_mod_map

Opt_D_dump_timings

Opt_D_dump_view_pattern_commoning

Opt_D_verbose_core2core

Opt_D_dump_debug

Opt_D_dump_json

Opt_D_ppr_debug

Opt_D_no_debug_output

Opt_D_dump_faststrings

data GeneralFlag

Enumerates the simple on-or-off dynamic flags

Constructors

Opt_DumpToFile
Append dump output to files instead of stdout.

Opt_D_faststring_stats

Opt_D_dump_minimal_imports

Opt_DoCoreLinting

Opt_DoLinearCoreLinting

Opt_DoStgLinting

Opt_DoCmmLinting

Opt_DoAsmLinting

Opt_DoAnnotationLinting

Opt_DoBoundsChecking

Opt_NoLlvmMangler

Opt_FastLlvm

Opt_NoTypeableBinds

Opt_DistinctConstructorTables

Opt_InfoTableMap

Opt_WarnIsError

Opt_ShowWarnGroups

Opt_HideSourcePaths

Opt_PrintExplicitForalls

Opt_PrintExplicitKinds

Opt_PrintExplicitCoercions

Opt_PrintExplicitRuntimeReps

Opt_PrintEqualityRelations

Opt_PrintAxiomIncomps

Opt_PrintUnicodeSyntax

Opt_PrintExpandedSynonyms

Opt_PrintPotentialInstances

Opt_PrintTypecheckerElaboration

Opt_CallArity

Opt_Exitification

Opt_Strictness

Opt_LateDmdAnal

Opt_KillAbsence

Opt_KillOneShot

Opt_FullLaziness

Opt_FloatIn

Opt_LateSpecialise

Opt_Specialise

Opt_SpecialiseAggressively

Opt_CrossModuleSpecialise

Opt_InlineGenerics

Opt_InlineGenericsAggressively

Opt_StaticArgumentTransformation

Opt_CSE

Opt_StgCSE

Opt_StgLiftLams

Opt_LiberateCase

Opt_SpecConstr

Opt_SpecConstrKeen

Opt_DoLambdaEtaExpansion

Opt_IgnoreAsserts

Opt_DoEtaReduction

Opt_CaseMerge

Opt_CaseFolding

Opt_UnboxStrictFields

Opt_UnboxSmallStrictFields

Opt_DictsCheap

Opt_EnableRewriteRules

Opt_EnableThSpliceWarnings

Opt_RegsGraph

Opt_RegsIterative

Opt_PedanticBottoms

Opt_LlvmTBAA

Opt_LlvmFillUndefWithGarbage

Opt_IrrefutableTuples

Opt_CmmSink

Opt_CmmStaticPred

Opt_CmmElimCommonBlocks

Opt_AsmShortcutting

Opt_OmitYields

Opt_FunToThunk

Opt_DictsStrict

Opt_DmdTxDictSel
deprecated, no effect and behaviour is now default. Allowed switching of a special demand transformer for dictionary selectors

Opt_Loopification

Opt_CfgBlocklayout
Use the cfg based block layout algorithm.

Opt_WeightlessBlocklayout
Layout based on last instruction per block.

Opt_CprAnal

Opt_WorkerWrapper

Opt_SolveConstantDicts

Opt_AlignmentSanitisation

Opt_CatchBottoms

Opt_NumConstantFolding

Opt_SimplPreInlining

Opt_IgnoreInterfacePragmas

Opt_OmitInterfacePragmas

Opt_ExposeAllUnfoldings

Opt_WriteInterface

Opt_WriteHie

Opt_AutoSccsOnIndividualCafs

Opt_ProfCountEntries

Opt_Pp

Opt_ForceRecomp

Opt_IgnoreOptimChanges

Opt_IgnoreHpcChanges

Opt_ExcessPrecision

Opt_EagerBlackHoling

Opt_NoHsMain

Opt_SplitSections

Opt_StgStats

Opt_HideAllPackages

Opt_HideAllPluginPackages

Opt_PrintBindResult

Opt_Haddock

Opt_HaddockOptions

Opt_BreakOnException

Opt_BreakOnError

Opt_PrintEvldWithShow

Opt_PrintBindContents

Opt_GenManifest

Opt_EmbedManifest

Opt_SharedImplib

Opt_BuildingCabalPackage

Opt_IgnoreDotGhci

Opt_GhciSandbox

Opt_GhciHistory

Opt_GhciLeakCheck

Opt_ValidateHie

Opt_LocalGhciHistory

Opt_NoIt

Opt_HelpfulErrors

Opt_DeferTypeErrors

Opt_DeferTypedHoles

Opt_DeferOutOfScopeVariables

Opt_PIC
-fPIC

Opt_PIE
-fPIE

Opt_PICExecutable
-pie

Opt_ExternalDynamicRefs

Opt_Ticky

Opt_Ticky_Allocd

Opt_Ticky_LNE

Opt_Ticky_Dyn_Thunk

Opt_RPath

Opt_RelativeDynlibPaths

Opt_CompactUnwind
-fcompact-unwind

Opt_Hpc

Opt_FamAppCache

Opt_ExternalInterpreter

Opt_OptimalApplicativeDo

Opt_VersionMacros

Opt_WholeArchiveHsLibs

Opt_SingleLibFolder

Opt_ExposeInternalSymbols

Opt_KeepCAFs

Opt_KeepGoing

Opt_ByteCode

Opt_LinkRts

Opt_ErrorSpans

Opt_DeferDiagnostics

Opt_DiagnosticsShowCaret

Opt_PprCaseAsLet

Opt_PprShowTicks

Opt_ShowHoleConstraints

Opt_ShowValidHoleFits

Opt_SortValidHoleFits

Opt_SortBySizeHoleFits

Opt_SortBySubsumHoleFits

Opt_AbstractRefHoleFits

Opt_UnclutterValidHoleFits

Opt_ShowTypeAppOfHoleFits

Opt_ShowTypeAppVarsOfHoleFits

Opt_ShowDocsOfHoleFits

Opt_ShowTypeOfHoleFits

Opt_ShowProvOfHoleFits

Opt_ShowMatchesOfHoleFits

Opt_ShowLoadedModules

Opt_HexWordLiterals

Opt_SuppressCoercions

Opt_SuppressVarKinds

Opt_SuppressModulePrefixes

Opt_SuppressTypeApplications

Opt_SuppressIdInfo

Opt_SuppressUnfoldings

Opt_SuppressTypeSignatures

Opt_SuppressUniques

Opt_SuppressStgExts

Opt_SuppressTicks

Opt_SuppressTimestamps
Suppress timestamps in dumps

Opt_AutoLinkPackages

Opt_ImplicitImportQualified

Opt_KeepHscppFiles

Opt_KeepHiDiffs

Opt_KeepHcFiles

Opt_KeepSFiles

Opt_KeepTmpFiles

Opt_KeepRawTokenStream

Opt_KeepLlvmFiles

Opt_KeepHiFiles

Opt_KeepOFiles

Opt_BuildDynamicToo

Opt_DistrustAllPackages

Opt_PackageTrust

Opt_PluginTrustworthy

Opt_G_NoStateHack

Opt_G_NoOptCoercion

data WarningFlag

Constructors

Opt_WarnDuplicateExports

Opt_WarnDuplicateConstraints

Opt_WarnRedundantConstraints

Opt_WarnHiShadows

Opt_WarnImplicitPrelude

Opt_WarnIncompletePatterns

Opt_WarnIncompleteUniPatterns

Opt_WarnIncompletePatternsRecUpd

Opt_WarnOverflowedLiterals

Opt_WarnEmptyEnumerations

Opt_WarnMissingFields

Opt_WarnMissingImportList

Opt_WarnMissingMethods

Opt_WarnMissingSignatures

Opt_WarnMissingLocalSignatures

Opt_WarnNameShadowing

Opt_WarnOverlappingPatterns

Opt_WarnTypeDefaults

Opt_WarnMonomorphism

Opt_WarnUnusedTopBinds

Opt_WarnUnusedLocalBinds

Opt_WarnUnusedPatternBinds

Opt_WarnUnusedImports

Opt_WarnUnusedMatches

Opt_WarnUnusedTypePatterns

Opt_WarnUnusedForalls

Opt_WarnUnusedRecordWildcards

Opt_WarnRedundantBangPatterns

Opt_WarnRedundantRecordWildcards

Opt_WarnWarningsDeprecations

Opt_WarnDeprecatedFlags

Opt_WarnMissingMonadFailInstances

Opt_WarnSemigroup

Opt_WarnDodgyExports

Opt_WarnDodgyImports

Opt_WarnOrphans

Opt_WarnAutoOrphans

Opt_WarnIdentities

Opt_WarnTabs

Opt_WarnUnrecognisedPragmas

Opt_WarnDodgyForeignImports

Opt_WarnUnusedDoBind

Opt_WarnWrongDoBind

Opt_WarnAlternativeLayoutRuleTransitional

Opt_WarnUnsafe

Opt_WarnSafe

Opt_WarnTrustworthySafe

Opt_WarnMissedSpecs

Opt_WarnAllMissedSpecs

Opt_WarnUnsupportedCallingConventions

Opt_WarnUnsupportedLlvmVersion

Opt_WarnMissedExtraSharedLib

Opt_WarnInlineRuleShadowing

Opt_WarnTypedHoles

Opt_WarnPartialTypeSignatures

Opt_WarnMissingExportedSignatures

Opt_WarnUntickedPromotedConstructors

Opt_WarnDerivingTypeable

Opt_WarnDeferredTypeErrors

Opt_WarnDeferredOutOfScopeVariables

Opt_WarnNonCanonicalMonadInstances

Opt_WarnNonCanonicalMonadFailInstances

Opt_WarnNonCanonicalMonoidInstances

Opt_WarnMissingPatternSynonymSignatures

Opt_WarnUnrecognisedWarningFlags

Opt_WarnSimplifiableClassConstraints

Opt_WarnCPPUndef

Opt_WarnUnbangedStrictPatterns

Opt_WarnMissingHomeModules

Opt_WarnPartialFields

Opt_WarnMissingExportList

Opt_WarnInaccessibleCode

Opt_WarnStarIsType

Opt_WarnStarBinder

Opt_WarnImplicitKindVars

Opt_WarnSpaceAfterBang

Opt_WarnMissingDerivingStrategies

Opt_WarnPrepositiveQualifiedModule

Opt_WarnUnusedPackages

Opt_WarnInferredSafeImports

Opt_WarnMissingSafeHaskellMode

Opt_WarnCompatUnqualifiedImports

Opt_WarnDerivingDefaults

Opt_WarnInvalidHaddock

Opt_WarnOperatorWhitespaceExtConflict

Opt_WarnOperatorWhitespace

Opt_WarnAmbiguousFields

Opt_WarnImplicitLift

Opt_WarnMissingKindSignatures

Opt_WarnUnicodeBidirectionalFormatCharacters

data WarnReason

Used when outputting warnings: if a reason is given, it is displayed. If a warning isn't controlled by a flag, this is made explicit at the point of use.

Constructors

NoReason

Reason !WarningFlag
Warning was enabled with the flag

ErrReason !(Maybe WarningFlag)
Warning was made an error because of -Werror or -Werror=WarningFlag

data Language

Constructors

Haskell98

Haskell2010

GHC2021

optimisationFlags :: EnumSet GeneralFlag