GHC.Driver.IncludeSpecs

Documentation

data IncludeSpecs

Used to differentiate the scope an include needs to apply to. We have to split the include paths to avoid accidentally forcing recursive includes since -I overrides the system search paths. See #14312.

Constructors

IncludeSpecs

Fields

includePathsQuote :: [String]

includePathsGlobal :: [String]

includePathsQuoteImplicit :: [String]

See Note [Implicit include paths]

addGlobalInclude :: IncludeSpecs -> [String] -> IncludeSpecs

Append to the list of includes a path that shall be included using `-I` when the C compiler is called. These paths override system search paths.

addQuoteInclude :: IncludeSpecs -> [String] -> IncludeSpecs

Append to the list of includes a path that shall be included using `-iquote` when the C compiler is called. These paths only apply when quoted includes are used. e.g. #include "foo.h"

addImplicitQuoteInclude :: IncludeSpecs -> [String] -> IncludeSpecs

These includes are not considered while fingerprinting the flags for iface | See Note [Implicit include paths]

flattenIncludes :: IncludeSpecs -> [String]

Concatenate and flatten the list of global and quoted includes returning just a flat list of paths.