GHC.Driver.MakeSem

Description

Implementation of a jobserver using system semaphores.

JSem: parallelism semaphore backed

runJSemAbstractSem

Arguments

:: SemaphoreName
the system semaphore to use

-> (AbstractSem -> IO a)
the operation to run which requires a semaphore

-> IO a

Implement an abstract semaphore using a semaphore Jobserver which queries the system semaphore of the given name for resources.

System semaphores

data Semaphore

A system semaphore (POSIX or Win32).

newtype SemaphoreName

Constructors

SemaphoreName

Fields

getSemaphoreName :: String

Abstract semaphores

data AbstractSem

Abstraction over the operations of a semaphore.

Constructors

AbstractSem

Fields

acquireSem :: IO ()

releaseSem :: IO ()

withAbstractSem :: AbstractSem -> IO b -> IO b