GHC.Driver.Monad

Ghc monad stuff

class (Functor m, ExceptionMonad m, HasDynFlags m, HasLogger m) => GhcMonad (m :: Type -> Type) where

A monad that has all the features needed by GHC API calls.

In short, a GHC monad

allows embedding of IO actions,

can log warnings,

allows handling of (extensible) exceptions, and

maintains a current session.

If you do not use Ghc or GhcT, make sure to call initGhcMonad before any call to the GHC API functions can occur.

Methods

getSession :: m HscEnv

setSession :: HscEnv -> m ()

newtype Ghc a

A minimal implementation of a GhcMonad. If you need a custom monad, e.g., to maintain additional state consider wrapping this monad or using GhcT.

Constructors

Ghc

Fields

unGhc :: Session -> IO a

newtype GhcT (m :: Type -> Type) a

A monad transformer to add GHC specific features to another monad.

Note that the wrapped monad must support IO and handling of exceptions.

Constructors

GhcT

Fields

unGhcT :: Session -> m a

liftGhcT :: m a -> GhcT m a

reflectGhc :: Ghc a -> Session -> IO a

Reflect a computation in the Ghc monad into the IO monad.

You can use this to call functions returning an action in the Ghc monad inside an IO action. This is needed for some (too restrictive) callback arguments of some library functions:

libFunc :: String -> (Int -> IO a) -> IO a
ghcFunc :: Int -> Ghc a

ghcFuncUsingLibFunc :: String -> Ghc a -> Ghc a
ghcFuncUsingLibFunc str =
  reifyGhc $ \s ->
    libFunc $ \i -> do
      reflectGhc (ghcFunc i) s

reifyGhc :: (Session -> IO a) -> Ghc a

getSessionDynFlags :: GhcMonad m => m DynFlags

Grabs the DynFlags from the Session

liftIO :: MonadIO m => IO a -> m a

data Session

The Session is a handle to the complete state of a compilation session. A compilation session consists of a set of modules constituting the current program or library, the context for interactive evaluation, and various caches.

Constructors

Session !(IORef HscEnv)

withSession :: GhcMonad m => (HscEnv -> m a) -> m a

Call the argument with the current session.

modifySession :: GhcMonad m => (HscEnv -> HscEnv) -> m ()

Set the current session to the result of applying the current session to the argument.

modifySessionM :: GhcMonad m => (HscEnv -> m HscEnv) -> m ()

Set the current session to the result of applying the current session to the argument.

withTempSession :: GhcMonad m => (HscEnv -> HscEnv) -> m a -> m a

Call an action with a temporarily modified Session.

Logger

modifyLogger :: GhcMonad m => (Logger -> Logger) -> m ()

Modify the logger

pushLogHookM :: GhcMonad m => (LogAction -> LogAction) -> m ()

Push a log hook on the stack

popLogHookM :: GhcMonad m => m ()

Pop a log hook from the stack

pushJsonLogHookM :: GhcMonad m => (LogJsonAction -> LogJsonAction) -> m ()

popJsonLogHookM :: GhcMonad m => m ()

putLogMsgM :: GhcMonad m => MessageClass -> SrcSpan -> SDoc -> m ()

Put a log message

putMsgM :: GhcMonad m => SDoc -> m ()

Put a log message

withTimingM :: GhcMonad m => SDoc -> (b -> ()) -> m b -> m b

Time an action

Diagnostics

logDiagnostics :: GhcMonad m => Messages GhcMessage -> m ()

A monad that allows logging of diagnostics.

printException :: (HasLogger m, MonadIO m, HasDynFlags m) => SourceError -> m ()

Print the all diagnostics in a SourceError. Useful inside exception handlers.

type WarnErrLogger = forall (m :: Type -> Type). (HasDynFlags m, MonadIO m, HasLogger m) => Maybe SourceError -> m ()

A function called to log warnings and errors.

defaultWarnErrLogger :: WarnErrLogger