GHC.Driver.Pipeline.Execute

Documentation

newtype HookedUse a

Constructors

HookedUse

Fields

runHookedUse :: (Hooks, PhaseHook) -> IO a

runPipeline :: Hooks -> HookedUse a -> IO a

The default mechanism to run a pipeline, see Note [The Pipeline Monad]

runPhase :: TPhase out -> IO out

Default interpretation of each phase, in terms of IO.

runLlvmManglePhase :: PipeEnv -> HscEnv -> FilePath -> IO [Char]

runMergeForeign :: PipeEnv -> HscEnv -> FilePath -> [FilePath] -> IO FilePath

runLlvmLlcPhase :: PipeEnv -> HscEnv -> FilePath -> IO FilePath

runLlvmOptPhase :: PipeEnv -> HscEnv -> FilePath -> IO FilePath

runAsPhase :: Bool -> PipeEnv -> HscEnv -> Maybe ModLocation -> FilePath -> IO FilePath

runJsPhase :: PipeEnv -> HscEnv -> Maybe ModLocation -> FilePath -> IO FilePath

Run the JS Backend postHsc phase.

runForeignJsPhase :: PipeEnv -> HscEnv -> Maybe ModLocation -> FilePath -> IO FilePath

Deal with foreign JS files (embed them into .o files)

applyAssemblerInfoGetter :: DefunctionalizedAssemblerInfoGetter -> Logger -> DynFlags -> Platform -> IO CompilerInfo

applyAssemblerProg :: DefunctionalizedAssemblerProg -> Logger -> DynFlags -> Platform -> [Option] -> IO ()

runCcPhase :: Phase -> PipeEnv -> HscEnv -> Maybe ModLocation -> FilePath -> IO FilePath

runHscBackendPhase :: PipeEnv -> HscEnv -> ModuleName -> HscSource -> ModLocation -> HscBackendAction -> IO ([FilePath], ModIface, HomeModLinkable, FilePath)

runUnlitPhase :: HscEnv -> FilePath -> FilePath -> IO FilePath

getFileArgs :: HscEnv -> FilePath -> IO (DynFlags, Messages PsMessage, [Warn])

runCppPhase :: HscEnv -> FilePath -> FilePath -> IO FilePath

runHscPhase :: PipeEnv -> HscEnv -> FilePath -> HscSource -> IO (HscEnv, ModSummary, HscRecompStatus)

mkOneShotModLocation :: PipeEnv -> DynFlags -> HscSource -> ModuleName -> IO ModLocation

Calculate the ModLocation from the provided DynFlags. This function is only used in one-shot mode and therefore takes into account the effect of -o/-ohi flags (which do nothing in --make mode)

runHscTcPhase :: HscEnv -> ModSummary -> IO (FrontendResult, Messages GhcMessage)

runHscPostTcPhase :: HscEnv -> ModSummary -> FrontendResult -> Messages GhcMessage -> Maybe Fingerprint -> IO HscBackendAction

runHsPpPhase :: HscEnv -> FilePath -> FilePath -> FilePath -> IO FilePath

phaseOutputFilenameNew

Arguments

:: Phase
The next phase

-> PipeEnv

-> HscEnv

-> Maybe ModLocation
A ModLocation, if we are compiling a Haskell source file

-> IO FilePath

getOutputFilename :: Logger -> TmpFs -> Phase -> PipelineOutput -> String -> DynFlags -> Phase -> Maybe ModLocation -> IO FilePath

Computes the next output filename for something in the compilation pipeline. This is controlled by several variables:

Phase: the last phase to be run (e.g. stopPhase). This is used to tell if we're in the last phase or not, because in that case flags like -o may be important.

PipelineOutput: is this intended to be a Temporary or Persistent build output? Temporary files just go in a fresh temporary name.

String: what was the basename of the original input file?

DynFlags: the obvious thing

Phase: the phase we want to determine the output filename of.

Maybe ModLocation: the ModLocation of the module we're compiling; this can be used to override the default output of an object file. (TODO: do we actually need this?)

llvmOptions

Arguments

:: LlvmConfig

-> DynFlags

-> [(String, String)]
pairs of (opt, llc) arguments

LLVM Options. These are flags to be passed to opt and llc, to ensure consistency we list them in pairs, so that they form groups.

hscPostBackendPhase :: HscSource -> Backend -> Phase

What phase to run after one of the backend code generators has run

compileStub :: HscEnv -> FilePath -> IO FilePath

joinObjectFiles :: HscEnv -> [FilePath] -> FilePath -> IO ()

See Note [Object merging].

getHCFilePackages :: FilePath -> IO [UnitId]

linkDynLibCheck :: Logger -> TmpFs -> DynFlags -> UnitEnv -> [String] -> [UnitId] -> IO ()

touchObjectFile :: Logger -> DynFlags -> FilePath -> IO ()