GHC.Driver.Pipeline.LogQueue

Documentation

data LogQueue

Each module is given a unique LogQueue to redirect compilation messages to. A Nothing value contains the result of compilation, and denotes the end of the message queue.

Constructors

LogQueue

Fields

logQueueId :: !Int

logQueueMessages :: !(IORef [Maybe (MessageClass, SrcSpan, SDoc, LogFlags)])

logQueueSemaphore :: !(MVar ())

newLogQueue :: Int -> IO LogQueue

finishLogQueue :: LogQueue -> IO ()

writeLogQueue :: LogQueue -> (MessageClass, SrcSpan, SDoc, LogFlags) -> IO ()

parLogAction :: LogQueue -> LogAction

data LogQueueQueue

Constructors

LogQueueQueue Int (IntMap LogQueue)

initLogQueue :: TVar LogQueueQueue -> LogQueue -> STM ()

allLogQueues :: LogQueueQueue -> [LogQueue]

Return all items in the queue in ascending order

newLogQueueQueue :: LogQueueQueue

logThread :: Int -> Int -> Logger -> TVar Bool -> TVar LogQueueQueue -> IO (IO ())