GHC.Driver.Ppr

Description

Printing related functions that depend on session state (DynFlags)

Documentation

showSDoc :: DynFlags -> SDoc -> String

Show a SDoc as a String with the default user style

showSDocUnsafe :: SDoc -> String

showSDocForUser :: DynFlags -> UnitState -> NamePprCtx -> SDoc -> String

Allows caller to specify the NamePprCtx to use

showPpr :: Outputable a => DynFlags -> a -> String

showPprUnsafe :: Outputable a => a -> String

printForUser :: DynFlags -> Handle -> NamePprCtx -> Depth -> SDoc -> IO ()