GHC.Hs

Documentation

module Language.Haskell.Syntax

module GHC.Hs.Binds

module GHC.Hs.Decls

module GHC.Hs.Expr

module GHC.Hs.ImpExp

module GHC.Hs.Lit

module GHC.Hs.Pat

module GHC.Hs.Type

module GHC.Hs.Utils

module GHC.Hs.Doc

module GHC.Hs.Extension

module GHC.Parser.Annotation

data Fixity

data HsModule

Haskell Module

All we actually declare here is the top-level structure for a module.

Constructors

HsModule

Fields

hsmodAnn :: EpAnn AnnsModule

hsmodLayout :: LayoutInfo

Layout info for the module. For incomplete modules (e.g. the output of parseHeader), it is NoLayoutInfo.

hsmodName :: Maybe (LocatedA ModuleName)

Nothing: "module X where" is omitted (in which case the next field is Nothing too)

hsmodExports :: Maybe (LocatedL [LIE GhcPs])

Export list

Nothing: export list omitted, so export everything

Just []: export nothing

Just [...]: as you would expect...

AnnKeywordIds : AnnOpen ,AnnClose

hsmodImports :: [LImportDecl GhcPs]

We snaffle interesting stuff out of the imported interfaces early on, adding that info to TyDecls/etc; so this list is often empty, downstream.

hsmodDecls :: [LHsDecl GhcPs]

Type, class, value, and interface signature decls

hsmodDeprecMessage :: Maybe (LocatedP WarningTxt)

reason/explanation for warning/deprecation of this module

AnnKeywordIds : AnnOpen ,AnnClose

hsmodHaddockModHeader :: Maybe LHsDocString

Haddock module info and description, unparsed

AnnKeywordIds : AnnOpen ,AnnClose

data AnnsModule

Constructors

AnnsModule

Fields

am_main :: [AddEpAnn]

am_decls :: AnnList

data HsParsedModule

Constructors

HsParsedModule

Fields

hpm_module :: Located HsModule

hpm_src_files :: [FilePath]

extra source files (e.g. from #includes). The lexer collects these from '# file line' pragmas, which the C preprocessor leaves behind. These files and their timestamps are stored in the .hi file, so that we can force recompilation if any of them change (#3589)