GHC.Hs.Binds

Documentation

module Language.Haskell.Syntax.Binds

coveredByNamespaceSpecifier :: NamespaceSpecifier -> NameSpace -> Bool

Check if namespace is covered by a namespace specifier: * NoNamespaceSpecifier covers both namespaces * TypeNamespaceSpecifier covers the type namespace only * DataNamespaceSpecifier covers the data namespace only

emptyLHsBinds :: forall (idL :: Pass) idR. LHsBindsLR (GhcPass idL) idR

emptyLocalBinds :: forall (a :: Pass) (b :: Pass). HsLocalBindsLR (GhcPass a) (GhcPass b)

emptyValBindsIn :: forall (a :: Pass) (b :: Pass). HsValBindsLR (GhcPass a) (GhcPass b)

emptyValBindsOut :: forall (a :: Pass) (b :: Pass). HsValBindsLR (GhcPass a) (GhcPass b)

eqEmptyLocalBinds :: HsLocalBindsLR a b -> Bool

extractSpecPragName :: SourceText -> String

Extracts the name for a SPECIALIZE instance pragma. In hsSigDoc, the src field of SpecInstSig signature contains the SourceText for a SPECIALIZE instance pragma of the form: "SourceText {-# SPECIALIZE"

Extraction ensures that all variants of the pragma name (with a Z or an S) are output exactly as used in the pragma.

getTcMultAnn :: HsMultAnn GhcTc -> Mult

hasSpecPrags :: TcSpecPrags -> Bool

hsSigDoc :: forall (p :: Pass). IsPass p => Sig (GhcPass p) -> SDoc

isDefaultMethod :: TcSpecPrags -> Bool

isEmptyIPBindsPR :: forall (p :: Pass). HsIPBinds (GhcPass p) -> Bool

isEmptyIPBindsTc :: HsIPBinds GhcTc -> Bool

isEmptyLHsBinds :: forall (idL :: Pass) idR. LHsBindsLR (GhcPass idL) idR -> Bool

isEmptyValBinds :: forall (a :: Pass) (b :: Pass). HsValBindsLR (GhcPass a) (GhcPass b) -> Bool

noSpecPrags :: TcSpecPrags

overlappingNamespaceSpecifiers :: NamespaceSpecifier -> NamespaceSpecifier -> Bool

Check if namespace specifiers overlap, i.e. if they are equal or if at least one of them doesn't specify a namespace

plusHsValBinds :: forall (a :: Pass). HsValBinds (GhcPass a) -> HsValBinds (GhcPass a) -> HsValBinds (GhcPass a)

pprDeclList :: [SDoc] -> SDoc

pprHsMultAnn :: forall (id :: Pass). OutputableBndrId id => HsMultAnn (GhcPass id) -> SDoc

pprLHsBinds :: forall (idL :: Pass) (idR :: Pass). (OutputableBndrId idL, OutputableBndrId idR) => LHsBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

pprLHsBindsForUser :: forall (idL :: Pass) (idR :: Pass) (id2 :: Pass). (OutputableBndrId idL, OutputableBndrId idR, OutputableBndrId id2) => LHsBindsLR (GhcPass idL) (GhcPass idR) -> [LSig (GhcPass id2)] -> [SDoc]

pprMinimalSig :: OutputableBndr name => LBooleanFormula (GenLocated l name) -> SDoc

pprSpec :: OutputableBndr id => id -> SDoc -> InlinePragma -> SDoc

pprTcSpecPrags :: TcSpecPrags -> SDoc

pprTicks :: SDoc -> SDoc -> SDoc

pprVarSig :: OutputableBndr id => [id] -> SDoc -> SDoc

ppr_monobind :: forall (idL :: Pass) (idR :: Pass). (OutputableBndrId idL, OutputableBndrId idR) => HsBindLR (GhcPass idL) (GhcPass idR) -> SDoc

ppr_sig :: forall (p :: Pass). OutputableBndrId p => Sig (GhcPass p) -> SDoc

pragBrackets :: SDoc -> SDoc

pragSrcBrackets :: SourceText -> String -> SDoc -> SDoc

Using SourceText in case the pragma was spelled differently or used mixed case

setTcMultAnn :: Mult -> HsMultAnn GhcRn -> HsMultAnn GhcTc

data ABExport

Abstraction Bindings Export

Constructors

ABE

Fields

abe_poly :: Id

Any INLINE pragma is attached to this Id

abe_mono :: Id

abe_wrap :: HsWrapper

See Note [ABExport wrapper] Shape: (forall abs_tvs. abs_ev_vars => abe_mono) ~ abe_poly

abe_prags :: TcSpecPrags

SPECIALISE pragmas

data AbsBinds

Typechecked, generalised bindings, used in the output to the type checker. See Note [AbsBinds].

Constructors

AbsBinds

Fields

abs_tvs :: [TyVar]

abs_ev_vars :: [EvVar]

Includes equality constraints

abs_exports :: [ABExport]

AbsBinds only gets used when idL = idR after renaming, but these need to be idL's for the collect... code in HsUtil to have the right type

abs_ev_binds :: [TcEvBinds]

Evidence bindings Why a list? See GHC.Tc.TyCl.Instance Note [Typechecking plan for instance declarations]

abs_binds :: LHsBinds GhcTc

Typechecked user bindings

abs_sig :: Bool

data ActivationAnn

Constructors

ActivationAnn

Fields

aa_openc :: EpToken "["

aa_closec :: EpToken "]"

aa_tilde :: Maybe (EpToken "~")

aa_val :: Maybe EpaLocation

data AnnPSB

Constructors

AnnPSB

Fields

ap_pattern :: EpToken "pattern"

ap_openc :: Maybe (EpToken "{")

ap_closec :: Maybe (EpToken "}")

ap_larrow :: Maybe (EpUniToken "<-" "\8592")

ap_equal :: Maybe (EpToken "=")

data AnnSig

Constructors

AnnSig

Fields

asDcolon :: EpUniToken "::" "\8759"

asPattern :: Maybe (EpToken "pattern")

asDefault :: Maybe (EpToken "default")

data AnnSpecSig

Constructors

AnnSpecSig

Fields

ass_open :: EpaLocation

ass_close :: EpToken "#-}"

ass_dcolon :: TokDcolon

ass_act :: ActivationAnn

newtype IdSig

A type signature in generated code, notably the code generated for record selectors. We simply record the desired Id itself, replete with its name, type and IdDetails. Otherwise it's just like a type signature: there should be an accompanying binding

Constructors

IdSig

Fields

unIdSig :: Id

type LTcSpecPrag = Located TcSpecPrag

Located Type checker Specification Pragmas

data NHsValBindsLR idL

Constructors

NValBinds [(RecFlag, LHsBinds idL)] [LSig GhcRn]

data NamespaceSpecifier

Optional namespace specifier for fixity signatures, WARNINIG and DEPRECATED pragmas.

Examples:

{-# WARNING in "x-partial" data Head "don't use this pattern synonym" #-} -- ↑ DataNamespaceSpecifier

{-# DEPRECATED type D "This type was deprecated" #-} -- ↑ TypeNamespaceSpecifier

infixr 6 data $ -- ↑ DataNamespaceSpecifier

Constructors

NoNamespaceSpecifier

TypeNamespaceSpecifier (EpToken "type")

DataNamespaceSpecifier (EpToken "data")

data TcSpecPrag

Type checker Specification Pragma

Constructors

SpecPrag Id HsWrapper InlinePragma
The Id to be specialised, a wrapper that specialises the polymorphic function, and inlining spec for the specialised function

data TcSpecPrags

Type checker Specialisation Pragmas

TcSpecPrags conveys SPECIALISE pragmas from the type checker to the desugarer

Constructors

IsDefaultMethod
Super-specialised: a default method should be macro-expanded at every call site

SpecPrags [LTcSpecPrag]

type family XVarBindGhc (pL :: Pass) (pR :: Pass) where ...

Equations

XVarBindGhc 'Typechecked 'Typechecked = NoExtField

XVarBindGhc _1 _2 = DataConCantHappen

Orphan instances

OutputableBndrId p => Outputable (FixitySig (GhcPass p))

Instance details

Methods

ppr :: FixitySig (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsIPBinds (GhcPass p))

Instance details

Methods

ppr :: HsIPBinds (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (IPBind (GhcPass p))

Instance details

Methods

ppr :: IPBind (GhcPass p) -> SDoc

Outputable (XRec pass (IdP pass)) => Outputable (RecordPatSynField pass)

Instance details

Methods

ppr :: RecordPatSynField pass -> SDoc

OutputableBndrId p => Outputable (Sig (GhcPass p))

Instance details

Methods

ppr :: Sig (GhcPass p) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr) => Outputable (HsBindLR (GhcPass pl) (GhcPass pr))

Instance details

Methods

ppr :: HsBindLR (GhcPass pl) (GhcPass pr) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr) => Outputable (HsLocalBindsLR (GhcPass pl) (GhcPass pr))

Instance details

Methods

ppr :: HsLocalBindsLR (GhcPass pl) (GhcPass pr) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr) => Outputable (HsValBindsLR (GhcPass pl) (GhcPass pr))

Instance details

Methods

ppr :: HsValBindsLR (GhcPass pl) (GhcPass pr) -> SDoc

(OutputableBndrId l, OutputableBndrId r) => Outputable (PatSynBind (GhcPass l) (GhcPass r))

Instance details

Methods

ppr :: PatSynBind (GhcPass l) (GhcPass r) -> SDoc