GHC.Hs.Doc

Description

Types and functions for raw and lexed docstrings.

Documentation

type HsDoc = WithHsDocIdentifiers HsDocString

A docstring with the (probable) identifiers found in it.

data WithHsDocIdentifiers a pass

Annotate a value with the probable identifiers found in it These will be used by haddock to generate links.

The identifiers are bundled along with their location in the source file. This is useful for tooling to know exactly where they originate.

This type is currently used in two places - for regular documentation comments, with a set to HsDocString, and for adding identifier information to warnings, where a is StringLiteral

Constructors

WithHsDocIdentifiers

Fields

hsDocString :: !a

hsDocIdentifiers :: ![Located (IdP pass)]

hsDocIds :: WithHsDocIdentifiers a GhcRn -> NameSet

Extract a mapping from the lexed identifiers to the names they may correspond to.

type LHsDoc pass = Located (HsDoc pass)

pprHsDocDebug :: Outputable (IdP name) => HsDoc name -> SDoc

Print a doc with its identifiers, useful for debugging

pprWithDoc :: LHsDoc name -> SDoc -> SDoc

Pretty print a thing with its doc The docstring will include the comment decorators '-- |', '{-|' etc and will come either before or after depending on how it was written i.e it will come after the thing if it is a '-- ^' or '{-^' and before otherwise.

pprMaybeWithDoc :: Maybe (LHsDoc name) -> SDoc -> SDoc

See pprWithHsDoc

module GHC.Hs.DocString

data ExtractedTHDocs

Maps of docs that were added via Template Haskell's putDoc.

Constructors

ExtractedTHDocs

Fields

ethd_mod_header :: Maybe (HsDoc GhcRn)

The added module header documentation, if it exists.

ethd_decl_docs :: UniqMap Name (HsDoc GhcRn)

The documentation added to declarations.

ethd_arg_docs :: UniqMap Name (IntMap (HsDoc GhcRn))

The documentation added to function arguments.

ethd_inst_docs :: UniqMap Name (HsDoc GhcRn)

The documentation added to class and family instances.

data DocStructureItem

A simplified version of IE.

Constructors

DsiSectionHeading !Int !(HsDoc GhcRn)

DsiDocChunk !(HsDoc GhcRn)

DsiNamedChunkRef !String

DsiExports !Avails

DsiModExport

Fields

!(NonEmpty ModuleName)

We might re-export avails from multiple modules with a single export declaration. E.g. when we have

module M (module X) where
import R0 as X
import R1 as X

!Avails

type DocStructure = [DocStructureItem]

data Docs

Constructors

Docs

Fields

docs_mod_hdr :: Maybe (HsDoc GhcRn)

Module header.

docs_decls :: UniqMap Name [HsDoc GhcRn]

Docs for declarations: functions, data types, instances, methods etc. A list because sometimes subsequent haddock comments can be combined into one

docs_args :: UniqMap Name (IntMap (HsDoc GhcRn))

Docs for arguments. E.g. function arguments, method arguments.

docs_structure :: DocStructure

docs_named_chunks :: Map String (HsDoc GhcRn)

Map from chunk name to content.

This map will be empty unless we have an explicit export list from which we can reference the chunks.

docs_haddock_opts :: Maybe String

Haddock options from OPTIONS_HADDOCK or from -haddock-opts.

docs_language :: Maybe Language

The Language used in the module, for example Haskell2010.

docs_extensions :: EnumSet Extension

The full set of language extensions used in the module.

emptyDocs :: Docs