GHC.Hs.DocString

Description

An exactprintable structure for docstrings

Documentation

type LHsDocString = Located HsDocString

data HsDocString

Haskell Documentation String

Rich structure to support exact printing The location around each chunk doesn't include the decorators

Constructors

MultiLineDocString !HsDocStringDecorator !(NonEmpty LHsDocStringChunk)
The first chunk is preceded by "-- decorator" and each following chunk is preceded by "--" Example: -- | This is a docstring for foo. It is the line with the decorator '|' and is always included -- This continues that docstring and is the second element in the NonEmpty list foo :: a -> a

NestedDocString !HsDocStringDecorator LHsDocStringChunk
The docstring is preceded by "{-decorator" and followed by "-}" The chunk contains balanced pairs of '{-' and '-}'

GeneratedDocString HsDocStringChunk
A docstring generated either internally or via TH Pretty printed with the '-- |' decorator This is because it may contain unbalanced pairs of '{-' and '-}' and not form a valid NestedDocString

data HsDocStringDecorator

Constructors

HsDocStringNext
'|' is the decorator

HsDocStringPrevious
^ is the decorator

HsDocStringNamed !String
'$string' is the decorator

HsDocStringGroup !Int
The decorator is the given number of *s

newtype HsDocStringChunk

A contiguous chunk of documentation

Constructors

HsDocStringChunk ByteString

type LHsDocStringChunk = Located HsDocStringChunk

isEmptyDocString :: HsDocString -> Bool

unpackHDSC :: HsDocStringChunk -> String

mkHsDocStringChunk :: String -> HsDocStringChunk

mkHsDocStringChunkUtf8ByteString :: ByteString -> HsDocStringChunk

Create a HsDocString from a UTF8-encoded ByteString.

pprHsDocString :: HsDocString -> SDoc

Pretty print with decorators, exactly as the user wrote it

pprHsDocStrings :: [HsDocString] -> SDoc

mkGeneratedHsDocString :: String -> HsDocString

docStringChunks :: HsDocString -> [LHsDocStringChunk]

renderHsDocString :: HsDocString -> String

Just get the docstring, without any decorators

renderHsDocStrings :: [HsDocString] -> String

Just get the docstring, without any decorators Separates docstrings using "nn", which is how haddock likes to render them

exactPrintHsDocString :: HsDocString -> String

Pretty print with decorators, exactly as the user wrote it

pprWithDocString :: HsDocString -> SDoc -> SDoc

Annotate a pretty printed thing with its doc The docstring comes after if is HsDocStringPrevious Otherwise it comes before. Note - we convert MultiLineDocString HsDocStringPrevious to HsDocStringNext because we can't control if something else will be pretty printed on the same line