GHC.Hs.Dump

Description

Contains a debug function to dump parts of the GHC.Hs AST. It uses a syb traversal which falls back to displaying based on the constructor name, so can be used to dump anything having a Data.Data instance.

Dumping ASTs

showAstData :: Data a => BlankSrcSpan -> BlankEpAnnotations -> a -> SDoc

Show a GHC syntax tree. This parameterised because it is also used for comparing ASTs in ppr roundtripping tests, where the SrcSpan's are blanked out, to avoid comparing locations, only structure

data BlankSrcSpan

Constructors

BlankSrcSpan

BlankSrcSpanFile

NoBlankSrcSpan

data BlankEpAnnotations

Constructors

BlankEpAnnotations

NoBlankEpAnnotations