GHC.Hs.Expr

Description

Abstract Haskell syntax for expressions.

Expressions proper

type LHsExpr p

Arguments

= Located (HsExpr p)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Expression

type PostTcExpr = HsExpr GhcTc

Post-Type checking Expression

PostTcExpr is an evidence expression attached to the syntax tree by the type checker (c.f. postTcType).

type PostTcTable = [(Name, PostTcExpr)]

Post-Type checking Table

We use a PostTcTable where there are a bunch of pieces of evidence, more than is convenient to keep individually.

type family SyntaxExpr p

Syntax Expression

SyntaxExpr is represents the function used in interpreting rebindable syntax. In the parser, we have no information to supply; in the renamer, we have the name of the function (but see Note [Monad fail : Rebindable syntax, overloaded strings] for a wrinkle) and in the type-checker we have a more elaborate structure SyntaxExprTc.

In some contexts, rebindable syntax is not implemented, and so we have constructors to represent that possibility in both the renamer and typechecker instantiations.

E.g. (>>=) is filled in before the renamer by the appropriate Name for (>>=), and then instantiated by the type checker with its type args etc

type family SyntaxExprGhc (p :: Pass) = (r :: Type) | r -> p where ...

Equations

SyntaxExprGhc 'Parsed = NoExtField

SyntaxExprGhc 'Renamed = SyntaxExprRn

SyntaxExprGhc 'Typechecked = SyntaxExprTc

data SyntaxExprRn

The function to use in rebindable syntax. See Note [NoSyntaxExpr].

Constructors

SyntaxExprRn (HsExpr GhcRn)

NoSyntaxExprRn

data SyntaxExprTc

An expression with wrappers, used for rebindable syntax

This should desugar to

syn_res_wrap $ syn_expr (syn_arg_wraps[0] arg0)
                        (syn_arg_wraps[1] arg1) ...

where the actual arguments come from elsewhere in the AST.

Constructors

SyntaxExprTc

Fields

syn_expr :: HsExpr GhcTc

syn_arg_wraps :: [HsWrapper]

syn_res_wrap :: HsWrapper

NoSyntaxExprTc

noExpr :: HsExpr (GhcPass p)

This is used for rebindable-syntax pieces that are too polymorphic for tcSyntaxOp (trS_fmap and the mzip in ParStmt)

noSyntaxExpr :: forall p. IsPass p => SyntaxExpr (GhcPass p)

mkSyntaxExpr :: HsExpr GhcRn -> SyntaxExprRn

Make a 'SyntaxExpr GhcRn' from an expression Used only in getMonadFailOp. See Note [Monad fail : Rebindable syntax, overloaded strings] in GHC.Rename.Expr

mkRnSyntaxExpr :: Name -> SyntaxExprRn

Make a SyntaxExpr from a Name (the "rn" is because this is used in the renamer).

type CmdSyntaxTable p = [(Name, HsExpr p)]

Command Syntax Table (for Arrow syntax)

data HsExpr p

A Haskell expression.

Constructors

HsVar (XVar p) (Located (IdP p))
Variable

HsUnboundVar (XUnboundVar p) OccName
Unbound variable; also used for "holes" (_ or _x). Turned from HsVar to HsUnboundVar by the renamer, when it finds an out-of-scope variable or hole. Turned into HsVar by type checker, to support deferred type errors.

HsConLikeOut (XConLikeOut p) ConLike
After typechecker only; must be different HsVar for pretty printing

HsRecFld (XRecFld p) (AmbiguousFieldOcc p)
Variable pointing to record selector Not in use after typechecking

HsOverLabel (XOverLabel p) (Maybe (IdP p)) FastString
Overloaded label (Note [Overloaded labels] in GHC.OverloadedLabels) Just id means RebindableSyntax is in use, and gives the id of the in-scope fromLabel. NB: Not in use after typechecking

HsIPVar (XIPVar p) HsIPName
Implicit parameter (not in use after typechecking)

HsOverLit (XOverLitE p) (HsOverLit p)
Overloaded literals

HsLit (XLitE p) (HsLit p)
Simple (non-overloaded) literals

HsLam (XLam p) (MatchGroup p (LHsExpr p))
Lambda abstraction. Currently always a single match

AnnKeywordId : AnnLam, AnnRarrow,

HsLamCase (XLamCase p) (MatchGroup p (LHsExpr p))
Lambda-case

AnnKeywordId : AnnLam, AnnCase,AnnOpen, AnnClose

HsApp (XApp p) (LHsExpr p) (LHsExpr p)
Application

HsAppType (XAppTypeE p) (LHsExpr p) (LHsWcType (NoGhcTc p))
Visible type application

Explicit type argument; e.g f @Int x y NB: Has wildcards, but no implicit quantification

AnnKeywordId : AnnAt,

OpApp (XOpApp p) (LHsExpr p) (LHsExpr p) (LHsExpr p)
Operator applications: NB Bracketed ops such as (+) come out as Vars.

NegApp (XNegApp p) (LHsExpr p) (SyntaxExpr p)
Negation operator. Contains the negated expression and the name of negate

AnnKeywordId : AnnMinus

HsPar

AnnKeywordId : AnnOpen '(', AnnClose ')'

Fields

(XPar p)

(LHsExpr p)

Parenthesised expr; see Note [Parens in HsSyn]

SectionL (XSectionL p) (LHsExpr p) (LHsExpr p)

SectionR (XSectionR p) (LHsExpr p) (LHsExpr p)

ExplicitTuple (XExplicitTuple p) [LHsTupArg p] Boxity
Used for explicit tuples and sections thereof

AnnKeywordId : AnnOpen, AnnClose

ExplicitSum (XExplicitSum p) ConTag Arity (LHsExpr p)
Used for unboxed sum types

AnnKeywordId : AnnOpen '(#', AnnVbar, AnnClose '#)',

There will be multiple AnnVbar, (1 - alternative) before the expression, (arity - alternative) after it

HsCase (XCase p) (LHsExpr p) (MatchGroup p (LHsExpr p))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsIf (XIf p) (LHsExpr p) (LHsExpr p) (LHsExpr p)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsMultiIf (XMultiIf p) [LGRHS p (LHsExpr p)]
Multi-way if

AnnKeywordId : AnnIf AnnOpen,AnnClose,

HsLet (XLet p) (LHsLocalBinds p) (LHsExpr p)
let(rec)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsDo (XDo p) (HsStmtContext GhcRn) (Located [ExprLStmt p])

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

ExplicitList (XExplicitList p) (Maybe (SyntaxExpr p)) [LHsExpr p]
Syntactic list: [a,b,c,...]

AnnKeywordId : AnnOpen '[', AnnClose ']'

RecordCon
Record construction

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rcon_ext :: XRecordCon p

rcon_con_name :: Located (IdP p)

rcon_flds :: HsRecordBinds p

RecordUpd
Record update

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rupd_ext :: XRecordUpd p

rupd_expr :: LHsExpr p

rupd_flds :: [LHsRecUpdField p]

ExprWithTySig (XExprWithTySig p) (LHsExpr p) (LHsSigWcType (NoGhcTc p))
Expression with an explicit type signature. e :: type

AnnKeywordId : AnnDcolon

ArithSeq (XArithSeq p) (Maybe (SyntaxExpr p)) (ArithSeqInfo p)
Arithmetic sequence

AnnKeywordId : AnnOpen '[', AnnComma,AnnDotdot, AnnClose ']'

HsBracket (XBracket p) (HsBracket p)

AnnKeywordId : AnnOpen, AnnOpenE,AnnOpenEQ, AnnClose,AnnCloseQ

HsRnBracketOut (XRnBracketOut p) (HsBracket GhcRn) [PendingRnSplice]

HsTcBracketOut (XTcBracketOut p) (Maybe QuoteWrapper) (HsBracket GhcRn) [PendingTcSplice]

HsSpliceE (XSpliceE p) (HsSplice p)

AnnKeywordId : AnnOpen, AnnClose

HsProc (XProc p) (LPat p) (LHsCmdTop p)
proc notation for Arrows

AnnKeywordId : AnnProc, AnnRarrow

HsStatic (XStatic p) (LHsExpr p)

AnnKeywordId : AnnStatic,

HsTick (XTick p) (Tickish (IdP p)) (LHsExpr p)

HsBinTick (XBinTick p) Int Int (LHsExpr p)

HsPragE (XPragE p) (HsPragE p) (LHsExpr p)

XExpr !(XXExpr p)

data RecordConTc

Extra data fields for a RecordCon, added by the type checker

Constructors

RecordConTc

Fields

rcon_con_like :: ConLike

rcon_con_expr :: PostTcExpr

data RecordUpdTc

Extra data fields for a RecordUpd, added by the type checker

Constructors

RecordUpdTc

Fields

rupd_cons :: [ConLike]

rupd_in_tys :: [Type]

rupd_out_tys :: [Type]

rupd_wrap :: HsWrapper

data HsWrap hs_syn

HsWrap appears only in typechecker output Invariant: The contained Expr is *NOT* itself an HsWrap. See Note [Detecting forced eta expansion] in GHC.HsToCore.Expr. This invariant is maintained by mkHsWrap. hs_syn is something like HsExpr or HsCmd

Constructors

HsWrap HsWrapper (hs_syn GhcTc)

data XXExprGhcTc

Constructors

WrapExpr !(HsWrap HsExpr)

ExpansionExpr !(HsExpansion (HsExpr GhcRn) (HsExpr GhcTc))

data HsExpansion a b

Constructors

HsExpanded a b

mkExpanded

Arguments

:: (HsExpansion a b -> b)
XExpr, XCmd, ...

-> a
source expression (GhcPs)

-> b
"desugared" expression (GhcRn)

-> b
suitably wrapped HsExpansion

Build a "wrapped" HsExpansion out of an extension constructor, and the two components of the expansion: original and desugared expressions.

See Note [Rebindable Syntax and HsExpansion] above for more details.

data HsPragE p

A pragma, written as {-# ... #-}, that may appear within an expression.

Constructors

HsPragSCC (XSCC p) SourceText StringLiteral

HsPragTick (XTickPragma p) SourceText (StringLiteral, (Int, Int), (Int, Int)) ((SourceText, SourceText), (SourceText, SourceText))

AnnKeywordId : AnnOpen, AnnOpen '{-# GENERATED', AnnVal,AnnVal, AnnColon,AnnVal, AnnMinus, AnnVal,AnnColon, AnnVal, AnnClose '#-}'

XHsPragE !(XXPragE p)

type LHsTupArg id = Located (HsTupArg id)

Located Haskell Tuple Argument

HsTupArg is used for tuple sections (,a,) is represented by ExplicitTuple [Missing ty1, Present a, Missing ty3] Which in turn stands for (x:ty1 y:ty2. (x,a,y))

data HsTupArg id

AnnKeywordId : AnnComma

Haskell Tuple Argument

Constructors

Present (XPresent id) (LHsExpr id)
The argument

Missing (XMissing id)
The argument is missing, but this is its type

XTupArg !(XXTupArg id)
Note [Trees that Grow] extension point

tupArgPresent :: LHsTupArg id -> Bool

pprLExpr :: OutputableBndrId p => LHsExpr (GhcPass p) -> SDoc

pprExpr :: OutputableBndrId p => HsExpr (GhcPass p) -> SDoc

isQuietHsExpr :: HsExpr id -> Bool

pprBinds :: (OutputableBndrId idL, OutputableBndrId idR) => HsLocalBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

ppr_lexpr :: OutputableBndrId p => LHsExpr (GhcPass p) -> SDoc

ppr_expr :: forall p. OutputableBndrId p => HsExpr (GhcPass p) -> SDoc

ppr_infix_expr :: forall p. OutputableBndrId p => HsExpr (GhcPass p) -> Maybe SDoc

ppr_apps :: OutputableBndrId p => HsExpr (GhcPass p) -> [Either (LHsExpr (GhcPass p)) (LHsWcType (NoGhcTc (GhcPass p)))] -> SDoc

pprExternalSrcLoc :: (StringLiteral, (Int, Int), (Int, Int)) -> SDoc

pprDebugParendExpr :: OutputableBndrId p => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprParendLExpr :: OutputableBndrId p => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprParendExpr :: OutputableBndrId p => PprPrec -> HsExpr (GhcPass p) -> SDoc

hsExprNeedsParens :: forall p. IsPass p => PprPrec -> HsExpr (GhcPass p) -> Bool

hsExprNeedsParens p e returns True if the expression e needs parentheses under precedence p.

parenthesizeHsExpr :: IsPass p => PprPrec -> LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

parenthesizeHsExpr p e checks if hsExprNeedsParens p e is true, and if so, surrounds e with an HsPar. Otherwise, it simply returns e.

stripParensLHsExpr :: LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

stripParensHsExpr :: HsExpr (GhcPass p) -> HsExpr (GhcPass p)

isAtomicHsExpr :: forall p. IsPass p => HsExpr (GhcPass p) -> Bool

type LHsCmd id = Located (HsCmd id)

Located Haskell Command (for arrow syntax)

data HsCmd id

Haskell Command (e.g. a "statement" in an Arrow proc block)

Constructors

HsCmdArrApp (XCmdArrApp id) (LHsExpr id) (LHsExpr id) HsArrAppType Bool

AnnKeywordId : Annlarrowtail, Annrarrowtail,AnnLarrowtail, AnnRarrowtail

HsCmdArrForm (XCmdArrForm id) (LHsExpr id) LexicalFixity (Maybe Fixity) [LHsCmdTop id]

AnnKeywordId : AnnOpenB '(|', AnnCloseB '|)'

HsCmdApp (XCmdApp id) (LHsCmd id) (LHsExpr id)

HsCmdLam (XCmdLam id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnLam, AnnRarrow,

HsCmdPar (XCmdPar id) (LHsCmd id)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsCmdCase (XCmdCase id) (LHsExpr id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsCmdLamCase (XCmdLamCase id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnLam, AnnCase,AnnOpen '{', AnnClose '}'

HsCmdIf (XCmdIf id) (SyntaxExpr id) (LHsExpr id) (LHsCmd id) (LHsCmd id)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsCmdLet (XCmdLet id) (LHsLocalBinds id) (LHsCmd id)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsCmdDo (XCmdDo id) (Located [CmdLStmt id])

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

XCmd !(XXCmd id)

data HsArrAppType

Haskell Array Application Type

Constructors

HsHigherOrderApp

HsFirstOrderApp

type LHsCmdTop p = Located (HsCmdTop p)

Top-level command, introducing a new arrow. This may occur inside a proc (where the stack is empty) or as an argument of a command-forming operator.

Located Haskell Top-level Command

data HsCmdTop p

Haskell Top-level Command

Constructors

HsCmdTop (XCmdTop p) (LHsCmd p)

XCmdTop !(XXCmdTop p)

data CmdTopTc

Constructors

CmdTopTc Type Type (CmdSyntaxTable GhcTc)

pprLCmd :: OutputableBndrId p => LHsCmd (GhcPass p) -> SDoc

pprCmd :: OutputableBndrId p => HsCmd (GhcPass p) -> SDoc

isQuietHsCmd :: HsCmd id -> Bool

ppr_lcmd :: OutputableBndrId p => LHsCmd (GhcPass p) -> SDoc

ppr_cmd :: forall p. OutputableBndrId p => HsCmd (GhcPass p) -> SDoc

pprCmdArg :: OutputableBndrId p => HsCmdTop (GhcPass p) -> SDoc

type HsRecordBinds p = HsRecFields p (LHsExpr p)

Haskell Record Bindings

data MatchGroup p body

Constructors

MG

Fields

mg_ext :: XMG p body

mg_alts :: Located [LMatch p body]

mg_origin :: Origin

XMatchGroup !(XXMatchGroup p body)

data MatchGroupTc

Constructors

MatchGroupTc

Fields

mg_arg_tys :: [Scaled Type]

mg_res_ty :: Type

type LMatch id body = Located (Match id body)

Located Match

May have AnnKeywordId : AnnSemi when in a list

data Match p body

Constructors

Match

Fields

m_ext :: XCMatch p body

m_ctxt :: HsMatchContext (NoGhcTc p)

m_pats :: [LPat p]

m_grhss :: GRHSs p body

XMatch !(XXMatch p body)

isInfixMatch :: Match id body -> Bool

isEmptyMatchGroup :: MatchGroup id body -> Bool

isSingletonMatchGroup :: [LMatch id body] -> Bool

Is there only one RHS in this list of matches?

matchGroupArity :: MatchGroup (GhcPass id) body -> Arity

hsLMatchPats :: LMatch (GhcPass id) body -> [LPat (GhcPass id)]

data GRHSs p body

Guarded Right-Hand Sides

GRHSs are used both for pattern bindings and for Matches

AnnKeywordId : AnnVbar, AnnEqual,AnnWhere, AnnOpen,AnnClose AnnRarrow,AnnSemi

Constructors

GRHSs

Fields

grhssExt :: XCGRHSs p body

grhssGRHSs :: [LGRHS p body]

Guarded RHSs

grhssLocalBinds :: LHsLocalBinds p

The where clause

XGRHSs !(XXGRHSs p body)

type LGRHS id body = Located (GRHS id body)

Located Guarded Right-Hand Side

data GRHS p body

Guarded Right Hand Side.

Constructors

GRHS (XCGRHS p body) [GuardLStmt p] body

XGRHS !(XXGRHS p body)

pprMatches :: (OutputableBndrId idR, Outputable body) => MatchGroup (GhcPass idR) body -> SDoc

pprFunBind :: (OutputableBndrId idR, Outputable body) => MatchGroup (GhcPass idR) body -> SDoc

pprPatBind :: forall bndr p body. (OutputableBndrId bndr, OutputableBndrId p, Outputable body) => LPat (GhcPass bndr) -> GRHSs (GhcPass p) body -> SDoc

pprMatch :: (OutputableBndrId idR, Outputable body) => Match (GhcPass idR) body -> SDoc

pprGRHSs :: (OutputableBndrId idR, Outputable body) => HsMatchContext passL -> GRHSs (GhcPass idR) body -> SDoc

pprGRHS :: (OutputableBndrId idR, Outputable body) => HsMatchContext passL -> GRHS (GhcPass idR) body -> SDoc

pp_rhs :: Outputable body => HsMatchContext passL -> body -> SDoc

type LStmt id body = Located (StmtLR id id body)

Located do block Statement

type LStmtLR idL idR body = Located (StmtLR idL idR body)

Located Statement with separate Left and Right id's

type Stmt id body = StmtLR id id body

do block Statement

type CmdLStmt id = LStmt id (LHsCmd id)

Command Located Statement

type CmdStmt id = Stmt id (LHsCmd id)

Command Statement

type ExprLStmt id = LStmt id (LHsExpr id)

Expression Located Statement

type ExprStmt id = Stmt id (LHsExpr id)

Expression Statement

type GuardLStmt id = LStmt id (LHsExpr id)

Guard Located Statement

type GuardStmt id = Stmt id (LHsExpr id)

Guard Statement

type GhciLStmt id = LStmt id (LHsExpr id)

Ghci Located Statement

type GhciStmt id = Stmt id (LHsExpr id)

Ghci Statement

data StmtLR idL idR body

API Annotations when in qualifier lists or guards - AnnKeywordId : AnnVbar, AnnComma,AnnThen, AnnBy,AnnBy, AnnGroup,AnnUsing

Constructors

LastStmt (XLastStmt idL idR body) body (Maybe Bool) (SyntaxExpr idR)

BindStmt

Fields

(XBindStmt idL idR body)

Post renaming has optional fail and bind / (>>=) operator. Post typechecking, also has multiplicity of the argument and the result type of the function passed to bind; that is, (P, S) in (>>=) :: Q -> (R # P -> S) -> T See Note [The type of bind in Stmts]

(LPat idL)

body

ApplicativeStmt (XApplicativeStmt idL idR body) [(SyntaxExpr idR, ApplicativeArg idL)] (Maybe (SyntaxExpr idR))
ApplicativeStmt represents an applicative expression built with <$> and <*>. It is generated by the renamer, and is desugared into the appropriate applicative expression by the desugarer, but it is intended to be invisible in error messages.

For full details, see Note [ApplicativeDo] in GHC.Rename.Expr

BodyStmt (XBodyStmt idL idR body) body (SyntaxExpr idR) (SyntaxExpr idR)

LetStmt (XLetStmt idL idR body) (LHsLocalBindsLR idL idR)

AnnKeywordId : AnnLet AnnOpen '{',AnnClose '}',

ParStmt (XParStmt idL idR body) [ParStmtBlock idL idR] (HsExpr idR) (SyntaxExpr idR)

TransStmt

Fields

trS_ext :: XTransStmt idL idR body

trS_form :: TransForm

trS_stmts :: [ExprLStmt idL]

trS_bndrs :: [(IdP idR, IdP idR)]

trS_using :: LHsExpr idR

trS_by :: Maybe (LHsExpr idR)

trS_ret :: SyntaxExpr idR

trS_bind :: SyntaxExpr idR

trS_fmap :: HsExpr idR

RecStmt

AnnKeywordId : AnnRec

Fields

recS_ext :: XRecStmt idL idR body

recS_stmts :: [LStmtLR idL idR body]

recS_later_ids :: [IdP idR]

recS_rec_ids :: [IdP idR]

recS_bind_fn :: SyntaxExpr idR

recS_ret_fn :: SyntaxExpr idR

recS_mfix_fn :: SyntaxExpr idR

XStmtLR !(XXStmtLR idL idR body)

data RecStmtTc

Constructors

RecStmtTc

Fields

recS_bind_ty :: Type

recS_later_rets :: [PostTcExpr]

recS_rec_rets :: [PostTcExpr]

recS_ret_ty :: Type

data XBindStmtRn

Constructors

XBindStmtRn

Fields

xbsrn_bindOp :: SyntaxExpr GhcRn

xbsrn_failOp :: FailOperator GhcRn

data XBindStmtTc

Constructors

XBindStmtTc

Fields

xbstc_bindOp :: SyntaxExpr GhcTc

xbstc_boundResultType :: Type

xbstc_boundResultMult :: Mult

xbstc_failOp :: FailOperator GhcTc

data TransForm

Constructors

ThenForm

GroupForm

data ParStmtBlock idL idR

Parenthesised Statement Block

Constructors

ParStmtBlock (XParStmtBlock idL idR) [ExprLStmt idL] [IdP idR] (SyntaxExpr idR)

XParStmtBlock !(XXParStmtBlock idL idR)

type FailOperator id = Maybe (SyntaxExpr id)

The fail operator

This is used for `.. <-` "bind statments" in do notation, including non-monadic "binds" in applicative.

The fail operator is 'Just expr' if it potentially fail monadically. if the pattern match cannot fail, or shouldn't fail monadically (regular incomplete pattern exception), it is Nothing.

See Note [Monad fail : Rebindable syntax, overloaded strings] for the type of expression in the Just case, and why it is so.

See Note [Failing pattern matches in Stmts] for which contexts for 'BindStmt's should use the monadic fail and which shouldn't.

data ApplicativeArg idL

Applicative Argument

Constructors

ApplicativeArgOne

Fields

xarg_app_arg_one :: XApplicativeArgOne idL

The fail operator, after renaming

The fail operator is needed if this is a BindStmt where the pattern can fail. E.g.: (Just a) <- stmt The fail operator will be invoked if the pattern match fails. It is also used for guards in MonadComprehensions. The fail operator is Nothing if the pattern match can't fail

app_arg_pattern :: LPat idL

arg_expr :: LHsExpr idL

is_body_stmt :: Bool

True = was a BodyStmt, False = was a BindStmt. See Note [Applicative BodyStmt]

ApplicativeArgMany

Fields

xarg_app_arg_many :: XApplicativeArgMany idL

app_stmts :: [ExprLStmt idL]

final_expr :: HsExpr idL

bv_pattern :: LPat idL

stmt_context :: HsStmtContext GhcRn

XApplicativeArg !(XXApplicativeArg idL)

pprStmt :: forall idL idR body. (OutputableBndrId idL, OutputableBndrId idR, Outputable body) => StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc

pprArg :: forall idL. OutputableBndrId idL => ApplicativeArg (GhcPass idL) -> SDoc

pprTransformStmt :: OutputableBndrId p => [IdP (GhcPass p)] -> LHsExpr (GhcPass p) -> Maybe (LHsExpr (GhcPass p)) -> SDoc

pprTransStmt :: Outputable body => Maybe body -> body -> TransForm -> SDoc

pprBy :: Outputable body => Maybe body -> SDoc

pprDo :: (OutputableBndrId p, Outputable body) => HsStmtContext any -> [LStmt (GhcPass p) body] -> SDoc

ppr_module_name_prefix :: Maybe ModuleName -> SDoc

ppr_do_stmts :: (OutputableBndrId idL, OutputableBndrId idR, Outputable body) => [LStmtLR (GhcPass idL) (GhcPass idR) body] -> SDoc

pprComp :: (OutputableBndrId p, Outputable body) => [LStmt (GhcPass p) body] -> SDoc

pprQuals :: (OutputableBndrId p, Outputable body) => [LStmt (GhcPass p) body] -> SDoc

data HsSplice id

Haskell Splice

Constructors

HsTypedSplice (XTypedSplice id) SpliceDecoration (IdP id) (LHsExpr id)

HsUntypedSplice (XUntypedSplice id) SpliceDecoration (IdP id) (LHsExpr id)

HsQuasiQuote (XQuasiQuote id) (IdP id) (IdP id) SrcSpan FastString

HsSpliced (XSpliced id) ThModFinalizers (HsSplicedThing id)

XSplice !(XXSplice id)

newtype HsSplicedT

Constructors

HsSplicedT DelayedSplice

data SpliceDecoration

A splice can appear with various decorations wrapped around it. This data type captures explicitly how it was originally written, for use in the pretty printer.

Constructors

DollarSplice
$splice or $$splice

BareSplice
bare splice

isTypedSplice :: HsSplice id -> Bool

newtype ThModFinalizers

Finalizers produced by a splice with addModFinalizer

See Note [Delaying modFinalizers in untyped splices] in GHC.Rename.Splice. For how this is used.

Constructors

ThModFinalizers [ForeignRef (Q ())]

data DelayedSplice

Constructors

DelayedSplice TcLclEnv (LHsExpr GhcRn) TcType (LHsExpr GhcTc)

data HsSplicedThing id

Haskell Spliced Thing

Values that can result from running a splice.

Constructors

HsSplicedExpr (HsExpr id)
Haskell Spliced Expression

HsSplicedTy (HsType id)
Haskell Spliced Type

HsSplicedPat (Pat id)
Haskell Spliced Pattern

type SplicePointName = Name

data PendingRnSplice

Pending Renamer Splice

Constructors

PendingRnSplice UntypedSpliceFlavour SplicePointName (LHsExpr GhcRn)

data UntypedSpliceFlavour

Constructors

UntypedExpSplice

UntypedPatSplice

UntypedTypeSplice

UntypedDeclSplice

data PendingTcSplice

Pending Type-checker Splice

Constructors

PendingTcSplice SplicePointName (LHsExpr GhcTc)

pprPendingSplice :: OutputableBndrId p => SplicePointName -> LHsExpr (GhcPass p) -> SDoc

pprSpliceDecl :: OutputableBndrId p => HsSplice (GhcPass p) -> SpliceExplicitFlag -> SDoc

ppr_splice_decl :: OutputableBndrId p => HsSplice (GhcPass p) -> SDoc

pprSplice :: forall p. OutputableBndrId p => HsSplice (GhcPass p) -> SDoc

ppr_quasi :: OutputableBndr p => p -> p -> FastString -> SDoc

ppr_splice :: OutputableBndrId p => SDoc -> IdP (GhcPass p) -> LHsExpr (GhcPass p) -> SDoc -> SDoc

data HsBracket p

Haskell Bracket

Constructors

ExpBr (XExpBr p) (LHsExpr p)

PatBr (XPatBr p) (LPat p)

DecBrL (XDecBrL p) [LHsDecl p]

DecBrG (XDecBrG p) (HsGroup p)

TypBr (XTypBr p) (LHsType p)

VarBr (XVarBr p) Bool (IdP p)

TExpBr (XTExpBr p) (LHsExpr p)

XBracket !(XXBracket p)

isTypedBracket :: HsBracket id -> Bool

pprHsBracket :: OutputableBndrId p => HsBracket (GhcPass p) -> SDoc

thBrackets :: SDoc -> SDoc -> SDoc

thTyBrackets :: SDoc -> SDoc

data ArithSeqInfo id

Arithmetic Sequence Information

Constructors

From (LHsExpr id)

FromThen (LHsExpr id) (LHsExpr id)

FromTo (LHsExpr id) (LHsExpr id)

FromThenTo (LHsExpr id) (LHsExpr id) (LHsExpr id)

pp_dotdot :: SDoc

data HsMatchContext p

Haskell Match Context

Context of a pattern match. This is more subtle than it would seem. See Note [Varieties of pattern matches].

Constructors

FunRhs

Fields

mc_fun :: LIdP p

function binder of f

mc_fixity :: LexicalFixity

fixing of f

mc_strictness :: SrcStrictness

was f banged? See Note [FunBind vs PatBind]

LambdaExpr
Patterns of a lambda

CaseAlt
Patterns and guards on a case alternative

IfAlt
Guards of a multi-way if alternative

ProcExpr
Patterns of a proc

PatBindRhs
A pattern binding eg [y] <- e = e

PatBindGuards
Guards of pattern bindings, e.g., (Just b) | Just _ <- x = e | otherwise = e'

RecUpd
Record update [used only in GHC.HsToCore.Expr to tell matchWrapper what sort of runtime error message to generate]

StmtCtxt (HsStmtContext p)
Pattern of a do-stmt, list comprehension, pattern guard, etc

ThPatSplice
A Template Haskell pattern splice

ThPatQuote
A Template Haskell pattern quotation [p| (a,b) |]

PatSyn
A pattern synonym declaration

isPatSynCtxt :: HsMatchContext p -> Bool

data HsStmtContext p

Haskell Statement Context.

Constructors

ListComp

MonadComp

DoExpr (Maybe ModuleName)

ModuleName.

do { ... }

MDoExpr (Maybe ModuleName)

ModuleName.

mdo { ... } ie recursive do-expression

ArrowExpr
do-notation in an arrow-command context

GhciStmtCtxt
A command-line Stmt in GHCi pat <- rhs

PatGuard (HsMatchContext p)
Pattern guard for specified thing

ParStmtCtxt (HsStmtContext p)
A branch of a parallel stmt

TransStmtCtxt (HsStmtContext p)
A branch of a transform stmt

qualifiedDoModuleName_maybe :: HsStmtContext p -> Maybe ModuleName

isComprehensionContext :: HsStmtContext id -> Bool

isMonadStmtContext :: HsStmtContext id -> Bool

Is this a monadic context?

isMonadCompContext :: HsStmtContext id -> Bool

matchSeparator :: HsMatchContext p -> SDoc

pprMatchContext :: Outputable (IdP p) => HsMatchContext p -> SDoc

pprMatchContextNoun :: Outputable (IdP id) => HsMatchContext id -> SDoc

pprAStmtContext :: Outputable (IdP id) => HsStmtContext id -> SDoc

pprStmtContext :: Outputable (IdP id) => HsStmtContext id -> SDoc

prependQualified :: Maybe ModuleName -> SDoc -> SDoc

matchContextErrString :: OutputableBndrId p => HsMatchContext (GhcPass p) -> SDoc

pprMatchInCtxt :: (OutputableBndrId idR, Outputable body) => Match (GhcPass idR) body -> SDoc

pprStmtInCtxt :: (OutputableBndrId idL, OutputableBndrId idR, Outputable body) => HsStmtContext (GhcPass idL) -> StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc