GHC.Hs.ImpExp

Documentation

module Language.Haskell.Syntax.ImpExp

data XImportDeclPass

Constructors

XImportDeclPass

Fields

ideclAnn :: EpAnn EpAnnImportDecl

ideclSourceText :: SourceText

ideclImplicit :: Bool

GHC generates an ImportDecl to represent the invisible `import Prelude` that appears in any file that omits `import Prelude`, setting this field to indicate that the import doesn't appear in the original source. True => implicit import (of Prelude)

data EpAnnImportDecl

Constructors

EpAnnImportDecl

Fields

importDeclAnnImport :: EpaLocation

importDeclAnnPragma :: Maybe (EpaLocation, EpaLocation)

importDeclAnnSafe :: Maybe EpaLocation

importDeclAnnQualified :: Maybe EpaLocation

importDeclAnnPackage :: Maybe EpaLocation

importDeclAnnAs :: Maybe EpaLocation

importDeclQualifiedStyle :: Maybe EpaLocation -> Maybe EpaLocation -> (Maybe EpaLocation, ImportDeclQualifiedStyle)

Given two possible located qualified tokens, compute a style (in a conforming Haskell program only one of the two can be not Nothing). This is called from GHC.Parser.

isImportDeclQualified :: ImportDeclQualifiedStyle -> Bool

Convenience function to answer the question if an import decl. is qualified.

simpleImportDecl :: ModuleName -> ImportDecl GhcPs

ieName :: IE (GhcPass p) -> IdP (GhcPass p)

ieWrappedName :: IEWrappedName (GhcPass p) -> IdP (GhcPass p)

ieNames :: IE (GhcPass p) -> [IdP (GhcPass p)]

ieWrappedLName :: IEWrappedName (GhcPass p) -> LIdP (GhcPass p)

lieWrappedName :: LIEWrappedName (GhcPass p) -> IdP (GhcPass p)

ieLWrappedName :: LIEWrappedName (GhcPass p) -> LIdP (GhcPass p)

replaceWrappedName :: IEWrappedName GhcPs -> IdP GhcRn -> IEWrappedName GhcRn

replaceLWrappedName :: LIEWrappedName GhcPs -> IdP GhcRn -> LIEWrappedName GhcRn

pprImpExp :: (HasOccName name, OutputableBndr name) => name -> SDoc

Orphan instances

Data (IEWrappedName GhcPs)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> IEWrappedName GhcPs -> c (IEWrappedName GhcPs)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (IEWrappedName GhcPs)

toConstr :: IEWrappedName GhcPs -> Constr

dataTypeOf :: IEWrappedName GhcPs -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (IEWrappedName GhcPs))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (IEWrappedName GhcPs))

gmapT :: (forall b. Data b => b -> b) -> IEWrappedName GhcPs -> IEWrappedName GhcPs

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> IEWrappedName GhcPs -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> IEWrappedName GhcPs -> r

gmapQ :: (forall d. Data d => d -> u) -> IEWrappedName GhcPs -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> IEWrappedName GhcPs -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> IEWrappedName GhcPs -> m (IEWrappedName GhcPs)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> IEWrappedName GhcPs -> m (IEWrappedName GhcPs)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> IEWrappedName GhcPs -> m (IEWrappedName GhcPs)

Data (IEWrappedName GhcRn)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> IEWrappedName GhcRn -> c (IEWrappedName GhcRn)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (IEWrappedName GhcRn)

toConstr :: IEWrappedName GhcRn -> Constr

dataTypeOf :: IEWrappedName GhcRn -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (IEWrappedName GhcRn))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (IEWrappedName GhcRn))

gmapT :: (forall b. Data b => b -> b) -> IEWrappedName GhcRn -> IEWrappedName GhcRn

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> IEWrappedName GhcRn -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> IEWrappedName GhcRn -> r

gmapQ :: (forall d. Data d => d -> u) -> IEWrappedName GhcRn -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> IEWrappedName GhcRn -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> IEWrappedName GhcRn -> m (IEWrappedName GhcRn)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> IEWrappedName GhcRn -> m (IEWrappedName GhcRn)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> IEWrappedName GhcRn -> m (IEWrappedName GhcRn)

Data (IEWrappedName GhcTc)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> IEWrappedName GhcTc -> c (IEWrappedName GhcTc)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (IEWrappedName GhcTc)

toConstr :: IEWrappedName GhcTc -> Constr

dataTypeOf :: IEWrappedName GhcTc -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (IEWrappedName GhcTc))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (IEWrappedName GhcTc))

gmapT :: (forall b. Data b => b -> b) -> IEWrappedName GhcTc -> IEWrappedName GhcTc

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> IEWrappedName GhcTc -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> IEWrappedName GhcTc -> r

gmapQ :: (forall d. Data d => d -> u) -> IEWrappedName GhcTc -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> IEWrappedName GhcTc -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> IEWrappedName GhcTc -> m (IEWrappedName GhcTc)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> IEWrappedName GhcTc -> m (IEWrappedName GhcTc)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> IEWrappedName GhcTc -> m (IEWrappedName GhcTc)

(HasOccName (IdP (GhcPass p)), OutputableBndrId p) => HasOccName (IEWrappedName (GhcPass p))

Instance details

Methods

occName :: IEWrappedName (GhcPass p) -> OccName

OutputableBndrId p => Outputable (IE (GhcPass p))

Instance details

Methods

ppr :: IE (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (IEWrappedName (GhcPass p))

Instance details

Methods

ppr :: IEWrappedName (GhcPass p) -> SDoc

(OutputableBndrId p, Outputable (Anno (IE (GhcPass p))), Outputable (ImportDeclPkgQual (GhcPass p))) => Outputable (ImportDecl (GhcPass p))

Instance details

Methods

ppr :: ImportDecl (GhcPass p) -> SDoc

OutputableBndrId p => OutputableBndr (IEWrappedName (GhcPass p))

Instance details

Methods

pprBndr :: BindingSite -> IEWrappedName (GhcPass p) -> SDoc

pprPrefixOcc :: IEWrappedName (GhcPass p) -> SDoc

pprInfixOcc :: IEWrappedName (GhcPass p) -> SDoc

bndrIsJoin_maybe :: IEWrappedName (GhcPass p) -> Maybe Int

Eq (IEWrappedName GhcPs)

Instance details

Methods

(==) :: IEWrappedName GhcPs -> IEWrappedName GhcPs -> Bool

(/=) :: IEWrappedName GhcPs -> IEWrappedName GhcPs -> Bool

Eq (IEWrappedName GhcRn)

Instance details

Methods

(==) :: IEWrappedName GhcRn -> IEWrappedName GhcRn -> Bool

(/=) :: IEWrappedName GhcRn -> IEWrappedName GhcRn -> Bool

Eq (IEWrappedName GhcTc)

Instance details

Methods

(==) :: IEWrappedName GhcTc -> IEWrappedName GhcTc -> Bool

(/=) :: IEWrappedName GhcTc -> IEWrappedName GhcTc -> Bool