GHC.Hs.Lit

Description

Source-language literals

Documentation

module Language.Haskell.Syntax.Lit

convertLit :: HsLit (GhcPass p1) -> HsLit (GhcPass p2)

Convert a literal from one index type to another

overLitType :: HsOverLit GhcTc -> Type

pmPprHsLit :: HsLit (GhcPass x) -> SDoc

pmPprHsLit pretty prints literals and is used when pretty printing pattern match warnings. All are printed the same (i.e., without hashes if they are primitive and not wrapped in constructors if they are boxed). This happens mainly for too reasons: * We do not want to expose their internal representation * The warnings become too messy

pp_st_suffix :: SourceText -> SDoc -> SDoc -> SDoc

data OverLitTc

Constructors

OverLitTc

Fields

ol_rebindable :: Bool

ol_type :: Type

Orphan instances

Outputable (HsLit (GhcPass p))

Instance details

Methods

ppr :: HsLit (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsOverLit (GhcPass p))

Instance details

Methods

ppr :: HsOverLit (GhcPass p) -> SDoc