GHC.Hs.Pat

Documentation

data Pat p

Pattern

AnnKeywordId : AnnBang

Constructors

WildPat (XWildPat p)
Wildcard Pattern The sole reason for a type on a WildPat is to support hsPatType :: Pat Id -> Type

VarPat (XVarPat p) (LIdP p)
Variable Pattern

LazyPat (XLazyPat p) (LPat p)
Lazy Pattern ^ - AnnKeywordId : AnnTilde

AsPat (XAsPat p) (LIdP p) (LPat p)
As pattern ^ - AnnKeywordId : AnnAt

ParPat (XParPat p) (LPat p)
Parenthesised pattern See Note [Parens in HsSyn] in GHC.Hs.Expr ^ - AnnKeywordId : AnnOpen '(', AnnClose ')'

BangPat (XBangPat p) (LPat p)
Bang pattern ^ - AnnKeywordId : AnnBang

ListPat (XListPat p) [LPat p]
Syntactic List

AnnKeywordId : AnnOpen '[', AnnClose ']'

TuplePat (XTuplePat p) [LPat p] Boxity
Tuple sub-patterns

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

SumPat (XSumPat p) (LPat p) ConTag Arity
Anonymous sum pattern

AnnKeywordId : AnnOpen '(#', AnnClose '#)'

ConPat
Constructor Pattern

Fields

pat_con_ext :: XConPat p

pat_con :: XRec p (ConLikeP p)

pat_args :: HsConPatDetails p

ViewPat

AnnKeywordId : AnnRarrow

Fields

(XViewPat p)

(LHsExpr p)

(LPat p)

View Pattern

SplicePat

AnnKeywordId : AnnOpen '$(' AnnClose ')'

Fields

(XSplicePat p)

(HsSplice p)

Splice Pattern (Includes quasi-quotes)

LitPat (XLitPat p) (HsLit p)
Literal Pattern Used for *non-overloaded* literal patterns: Int#, Char#, Int, Char, String, etc.

NPat (XNPat p) (XRec p (HsOverLit p)) (Maybe (SyntaxExpr p)) (SyntaxExpr p)
Natural Pattern

AnnKeywordId : AnnVal +

NPlusKPat (XNPlusKPat p) (LIdP p) (XRec p (HsOverLit p)) (HsOverLit p) (SyntaxExpr p) (SyntaxExpr p)
n+k pattern

SigPat

AnnKeywordId : AnnDcolon

Fields

(XSigPat p)

(LPat p)

(HsPatSigType (NoGhcTc p))

Pattern with a type signature

XPat !(XXPat p)
Trees that Grow extension point for new constructors

type LPat p = XRec p (Pat p)

data EpAnnSumPat

Constructors

EpAnnSumPat

Fields

sumPatParens :: [AddEpAnn]

sumPatVbarsBefore :: [EpaLocation]

sumPatVbarsAfter :: [EpaLocation]

data ConPatTc

This is the extension field for ConPat, added after typechecking It adds quite a few extra fields, to support elaboration of pattern matching.

Constructors

ConPatTc

Fields

cpt_arg_tys :: [Type]

The universal arg types 1-1 with the universal tyvars of the constructor/pattern synonym Use (conLikeResTy pat_con cpt_arg_tys) to get the type of the pattern

cpt_tvs :: [TyVar]

Existentially bound type variables in correctly-scoped order e.g. [k:* x:k]

cpt_dicts :: [EvVar]

Ditto *coercion variables* and *dictionaries* One reason for putting coercion variable here I think is to ensure their kinds are zonked

cpt_binds :: TcEvBinds

Bindings involving those dictionaries

cpt_wrap :: HsWrapper

data CoPat

Coercion Pattern (translation only)

During desugaring a (CoPat co pat) turns into a cast with co on the scrutinee, followed by a match on pat.

Constructors

CoPat

Fields

co_cpt_wrap :: HsWrapper

Coercion Pattern If co :: t1 ~ t2, p :: t2, then (CoPat co p) :: t1

co_pat_inner :: Pat GhcTc

Why not LPat? Ans: existing locn will do

co_pat_ty :: Type

Type of whole pattern, t1

data ListPatTc

Constructors

ListPatTc Type (Maybe (Type, SyntaxExpr GhcTc))

type family ConLikeP x

type HsConPatDetails p = HsConDetails (HsPatSigType (NoGhcTc p)) (LPat p) (HsRecFields p (LPat p))

Haskell Constructor Pattern Details

hsConPatArgs :: forall p. UnXRec p => HsConPatDetails p -> [LPat p]

data HsRecFields p arg

Haskell Record Fields

HsRecFields is used only for patterns and expressions (not data type declarations)

Constructors

HsRecFields

Fields

rec_flds :: [LHsRecField p arg]

rec_dotdot :: Maybe (Located Int)

data HsRecField' id arg

Haskell Record Field

AnnKeywordId : AnnEqual,

For details on above see note [exact print annotations] in GHC.Parser.Annotation

Constructors

HsRecField

Fields

hsRecFieldAnn :: XHsRecField id

hsRecFieldLbl :: Located id

hsRecFieldArg :: arg

Filled in by renamer when punning

hsRecPun :: Bool

Note [Punning]

type LHsRecField' p id arg = XRec p (HsRecField' id arg)

Located Haskell Record Field

type HsRecField p arg = HsRecField' (FieldOcc p) arg

Haskell Record Field

type LHsRecField p arg = XRec p (HsRecField p arg)

Located Haskell Record Field

type HsRecUpdField p = HsRecField' (AmbiguousFieldOcc p) (LHsExpr p)

Haskell Record Update Field

type LHsRecUpdField p = XRec p (HsRecUpdField p)

Located Haskell Record Update Field

hsRecFields :: forall p arg. UnXRec p => HsRecFields p arg -> [XCFieldOcc p]

hsRecFieldSel :: HsRecField pass arg -> Located (XCFieldOcc pass)

hsRecFieldId :: HsRecField GhcTc arg -> Located Id

hsRecFieldsArgs :: forall p arg. UnXRec p => HsRecFields p arg -> [arg]

hsRecUpdFieldId :: HsRecField' (AmbiguousFieldOcc GhcTc) arg -> Located Id

hsRecUpdFieldOcc :: HsRecField' (AmbiguousFieldOcc GhcTc) arg -> LFieldOcc GhcTc

hsRecUpdFieldRdr :: HsRecUpdField (GhcPass p) -> Located RdrName

mkPrefixConPat :: DataCon -> [LPat GhcTc] -> [Type] -> LPat GhcTc

mkCharLitPat :: SourceText -> Char -> LPat GhcTc

mkNilPat :: Type -> LPat GhcTc

isSimplePat :: LPat (GhcPass x) -> Maybe (IdP (GhcPass x))

Is the pattern any of combination of:

(pat)

pat :: Type

~pat

!pat

x (variable)

looksLazyPatBind :: HsBind (GhcPass p) -> Bool

isBangedLPat :: LPat (GhcPass p) -> Bool

patNeedsParens :: forall p. IsPass p => PprPrec -> Pat (GhcPass p) -> Bool

patNeedsParens p pat returns True if the pattern pat needs parentheses under precedence p.

parenthesizePat :: IsPass p => PprPrec -> LPat (GhcPass p) -> LPat (GhcPass p)

parenthesizePat p pat checks if patNeedsParens p pat is true, and if so, surrounds pat with a ParPat. Otherwise, it simply returns pat.

isIrrefutableHsPat :: forall p. OutputableBndrId p => DynFlags -> LPat (GhcPass p) -> Bool

collectEvVarsPat :: Pat GhcTc -> Bag EvVar

collectEvVarsPats :: [Pat GhcTc] -> Bag EvVar

pprParendLPat :: OutputableBndrId p => PprPrec -> LPat (GhcPass p) -> SDoc

pprConArgs :: (OutputableBndrId p, Outputable (Anno (IdGhcP p))) => HsConPatDetails (GhcPass p) -> SDoc

pprLPat :: OutputableBndrId p => LPat (GhcPass p) -> SDoc

Orphan instances

OutputableBndrId p => Outputable (Pat (GhcPass p))

Instance details

Methods

ppr :: Pat (GhcPass p) -> SDoc