GHC.Hs.Syn.Type

Description

Compute the Type of an HsExpr GhcPass in a pure fashion.

Note that this does not currently support the use case of annotating every subexpression in an HsExpr with its Type. For more information on this task, see #12706, #15320, #16804, and #17331.

Extracting types from HsExpr

lhsExprType :: LHsExpr GhcTc -> Type

Compute the Type of an LHsExpr GhcPass in a pure fashion.

hsExprType :: HsExpr GhcTc -> Type

Compute the Type of an HsExpr GhcPass in a pure fashion.

hsWrapperType :: HsWrapper -> Type -> Type

Extracting types from HsSyn

hsLitType :: HsLit (GhcPass p) -> Type

hsPatType :: Pat GhcTc -> Type

hsLPatType :: LPat GhcTc -> Type