GHC.Hs.Types

Documentation

data HsType pass

Haskell Type

Constructors

HsForAllTy

AnnKeywordId : AnnForall, AnnDot,AnnDarrow For details on above see note [Api annotations] in ApiAnnotation

Fields

hst_xforall :: XForAllTy pass

hst_fvf :: ForallVisFlag

hst_bndrs :: [LHsTyVarBndr pass]

hst_body :: LHsType pass

HsQualTy

Fields

hst_xqual :: XQualTy pass

hst_ctxt :: LHsContext pass

hst_body :: LHsType pass

HsTyVar (XTyVar pass) PromotionFlag (Located (IdP pass))

AnnKeywordId : None

HsAppTy (XAppTy pass) (LHsType pass) (LHsType pass)

AnnKeywordId : None

HsAppKindTy (XAppKindTy pass) (LHsType pass) (LHsKind pass)

HsFunTy (XFunTy pass) (LHsType pass) (LHsType pass)

AnnKeywordId : AnnRarrow,

HsListTy (XListTy pass) (LHsType pass)

AnnKeywordId : AnnOpen '[', AnnClose ']'

HsTupleTy (XTupleTy pass) HsTupleSort [LHsType pass]

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

HsSumTy (XSumTy pass) [LHsType pass]

AnnKeywordId : AnnOpen '(#', AnnClose '#)'@

HsOpTy (XOpTy pass) (LHsType pass) (Located (IdP pass)) (LHsType pass)

AnnKeywordId : None

HsParTy (XParTy pass) (LHsType pass)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsIParamTy (XIParamTy pass) (Located HsIPName) (LHsType pass)
(?x :: ty)

AnnKeywordId : AnnDcolon

HsStarTy (XStarTy pass) Bool

AnnKeywordId : None

HsKindSig (XKindSig pass) (LHsType pass) (LHsKind pass)
(ty :: kind)

AnnKeywordId : AnnOpen '(', AnnDcolon,AnnClose ')'

HsSpliceTy (XSpliceTy pass) (HsSplice pass)

AnnKeywordId : AnnOpen '$(', AnnClose ')'

HsDocTy (XDocTy pass) (LHsType pass) LHsDocString

AnnKeywordId : None

HsBangTy (XBangTy pass) HsSrcBang (LHsType pass)

AnnKeywordId : AnnOpen '{-# UNPACK' or '{-# NOUNPACK', AnnClose '#-}' AnnBang '!'

HsRecTy (XRecTy pass) [LConDeclField pass]

AnnKeywordId : AnnOpen '{', AnnClose '}'

HsExplicitListTy (XExplicitListTy pass) PromotionFlag [LHsType pass]

AnnKeywordId : AnnOpen "'[", AnnClose ']'

HsExplicitTupleTy (XExplicitTupleTy pass) [LHsType pass]

AnnKeywordId : AnnOpen "'(", AnnClose ')'

HsTyLit (XTyLit pass) HsTyLit

AnnKeywordId : None

HsWildCardTy (XWildCardTy pass)

AnnKeywordId : None

XHsType (XXType pass)

data NewHsTypeX

Constructors

NHsCoreTy Type

type LHsType pass

Arguments

= Located (HsType pass)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Type

type HsKind pass = HsType pass

Haskell Kind

type LHsKind pass

Arguments

= Located (HsKind pass)
AnnKeywordId : AnnDcolon

Located Haskell Kind

data HsTyVarBndr pass

Haskell Type Variable Binder

Constructors

UserTyVar (XUserTyVar pass) (Located (IdP pass))

KindedTyVar (XKindedTyVar pass) (Located (IdP pass)) (LHsKind pass)

AnnKeywordId : AnnOpen, AnnDcolon, AnnClose

XTyVarBndr (XXTyVarBndr pass)

type LHsTyVarBndr pass = Located (HsTyVarBndr pass)

Located Haskell Type Variable Binder

data ForallVisFlag

Is a forall invisible (e.g., forall a b. {...}, with a dot) or visible (e.g., forall a b -> {...}, with an arrow)?

Constructors

ForallVis
A visible forall (with an arrow)

ForallInvis
An invisible forall (with a dot)

data LHsQTyVars pass

Located Haskell Quantified Type Variables

Constructors

HsQTvs

Fields

hsq_ext :: XHsQTvs pass

hsq_explicit :: [LHsTyVarBndr pass]

XLHsQTyVars (XXLHsQTyVars pass)

data HsImplicitBndrs pass thing

Haskell Implicit Binders

Constructors

HsIB

Fields

hsib_ext :: XHsIB pass thing

hsib_body :: thing

XHsImplicitBndrs (XXHsImplicitBndrs pass thing)

data HsWildCardBndrs pass thing

Haskell Wildcard Binders

Constructors

HsWC

Fields

hswc_ext :: XHsWC pass thing

hswc_body :: thing

XHsWildCardBndrs (XXHsWildCardBndrs pass thing)

type LHsSigType pass = HsImplicitBndrs pass (LHsType pass)

Located Haskell Signature Type

type LHsSigWcType pass = HsWildCardBndrs pass (LHsSigType pass)

Located Haskell Signature Wildcard Type

type LHsWcType pass = HsWildCardBndrs pass (LHsType pass)

Located Haskell Wildcard Type

data HsTupleSort

Haskell Tuple Sort

Constructors

HsUnboxedTuple

HsBoxedTuple

HsConstraintTuple

HsBoxedOrConstraintTuple

type HsContext pass = [LHsType pass]

Haskell Context

type LHsContext pass

Arguments

= Located (HsContext pass)
AnnKeywordId : AnnUnit For details on above see note [Api annotations] in ApiAnnotation

Located Haskell Context

noLHsContext :: LHsContext pass

data HsTyLit

Haskell Type Literal

Constructors

HsNumTy SourceText Integer

HsStrTy SourceText FastString

newtype HsIPName

These names are used early on to store the names of implicit parameters. They completely disappear after type-checking.

Constructors

HsIPName FastString

hsIPNameFS :: HsIPName -> FastString

data HsArg tm ty

Constructors

HsValArg tm

HsTypeArg SrcSpan ty

HsArgPar SrcSpan

numVisibleArgs :: [HsArg tm ty] -> Arity

type LHsTypeArg p = HsArg (LHsType p) (LHsKind p)

type LBangType pass = Located (BangType pass)

Located Bang Type

type BangType pass = HsType pass

Bang Type

In the parser, strictness and packedness annotations bind more tightly than docstrings. This means that when consuming a BangType (and looking for HsBangTy) we must be ready to peer behind a potential layer of HsDocTy. See #15206 for motivation and getBangType for an example.

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as the user wrote them in the source code.

(HsSrcBang _ SrcUnpack SrcLazy) and (HsSrcBang _ SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsSrcBang _ NoSrcUnpack SrcLazy)

Constructors

HsSrcBang SourceText SrcUnpackedness SrcStrictness

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field, or one with an unlifted type

HsStrict
Strict but not unpacked field

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie '~'

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{--} specified

SrcNoUnpack
{--} specified

NoSrcUnpack
no unpack pragma

getBangType :: LHsType a -> LHsType a

getBangStrictness :: LHsType a -> HsSrcBang

data ConDeclField pass

Constructor Declaration Field

Constructors

ConDeclField

AnnKeywordId : AnnDcolon

Fields

cd_fld_ext :: XConDeclField pass

cd_fld_names :: [LFieldOcc pass]

See Note [ConDeclField passs]

cd_fld_type :: LBangType pass

cd_fld_doc :: Maybe LHsDocString

XConDeclField (XXConDeclField pass)

type LConDeclField pass

Arguments

= Located (ConDeclField pass)
May have AnnKeywordId : AnnComma when in a list

Located Constructor Declaration Field

pprConDeclFields :: OutputableBndrId p => [LConDeclField (GhcPass p)] -> SDoc

data HsConDetails arg rec

Haskell Constructor Details

Constructors

PrefixCon [arg]

RecCon rec

InfixCon arg arg

data FieldOcc pass

Field Occurrence

Represents an *occurrence* of an unambiguous field. We store both the RdrName the user originally wrote, and after the renamer, the selector function.

Constructors

FieldOcc

Fields

extFieldOcc :: XCFieldOcc pass

rdrNameFieldOcc :: Located RdrName

See Note [Located RdrNames] in GHC.Hs.Expr

XFieldOcc (XXFieldOcc pass)

type LFieldOcc pass = Located (FieldOcc pass)

Located Field Occurrence

mkFieldOcc :: Located RdrName -> FieldOcc GhcPs

data AmbiguousFieldOcc pass

Ambiguous Field Occurrence

Represents an *occurrence* of a field that is potentially ambiguous after the renamer, with the ambiguity resolved by the typechecker. We always store the RdrName that the user originally wrote, and store the selector function after the renamer (for unambiguous occurrences) or the typechecker (for ambiguous occurrences).

See Note [HsRecField and HsRecUpdField] in GHC.Hs.Pat and Note [Disambiguating record fields] in TcExpr. See Note [Located RdrNames] in GHC.Hs.Expr

Constructors

Unambiguous (XUnambiguous pass) (Located RdrName)

Ambiguous (XAmbiguous pass) (Located RdrName)

XAmbiguousFieldOcc (XXAmbiguousFieldOcc pass)

mkAmbiguousFieldOcc :: Located RdrName -> AmbiguousFieldOcc GhcPs

rdrNameAmbiguousFieldOcc :: AmbiguousFieldOcc (GhcPass p) -> RdrName

selectorAmbiguousFieldOcc :: AmbiguousFieldOcc GhcTc -> Id

unambiguousFieldOcc :: AmbiguousFieldOcc GhcTc -> FieldOcc GhcTc

ambiguousFieldOcc :: FieldOcc GhcTc -> AmbiguousFieldOcc GhcTc

mkAnonWildCardTy :: HsType GhcPs

pprAnonWildCard :: SDoc

mkHsImplicitBndrs :: thing -> HsImplicitBndrs GhcPs thing

mkHsWildCardBndrs :: thing -> HsWildCardBndrs GhcPs thing

hsImplicitBody :: HsImplicitBndrs (GhcPass p) thing -> thing

mkEmptyImplicitBndrs :: thing -> HsImplicitBndrs GhcRn thing

mkEmptyWildCardBndrs :: thing -> HsWildCardBndrs GhcRn thing

mkHsQTvs :: [LHsTyVarBndr GhcPs] -> LHsQTyVars GhcPs

hsQTvExplicit :: LHsQTyVars pass -> [LHsTyVarBndr pass]

emptyLHsQTvs :: LHsQTyVars GhcRn

isEmptyLHsQTvs :: LHsQTyVars GhcRn -> Bool

isHsKindedTyVar :: HsTyVarBndr pass -> Bool

Does this HsTyVarBndr come with an explicit kind annotation?

hsTvbAllKinded :: LHsQTyVars pass -> Bool

Do all type variables in this LHsQTyVars come with kind annotations?

isLHsForAllTy :: LHsType p -> Bool

hsScopedTvs :: LHsSigType GhcRn -> [Name]

hsWcScopedTvs :: LHsSigWcType GhcRn -> [Name]

dropWildCards :: LHsSigWcType pass -> LHsSigType pass

hsTyVarName :: HsTyVarBndr (GhcPass p) -> IdP (GhcPass p)

hsAllLTyVarNames :: LHsQTyVars GhcRn -> [Name]

hsLTyVarLocNames :: LHsQTyVars (GhcPass p) -> [Located (IdP (GhcPass p))]

hsLTyVarName :: LHsTyVarBndr (GhcPass p) -> IdP (GhcPass p)

hsLTyVarNames :: [LHsTyVarBndr (GhcPass p)] -> [IdP (GhcPass p)]

hsLTyVarLocName :: LHsTyVarBndr (GhcPass p) -> Located (IdP (GhcPass p))

hsExplicitLTyVarNames :: LHsQTyVars (GhcPass p) -> [IdP (GhcPass p)]

splitLHsInstDeclTy :: LHsSigType GhcRn -> ([Name], LHsContext GhcRn, LHsType GhcRn)

Decompose a type class instance type (of the form forall tvs. context => instance_head) into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (forall tvs. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

getLHsInstDeclHead :: LHsSigType (GhcPass p) -> LHsType (GhcPass p)

getLHsInstDeclClass_maybe :: LHsSigType (GhcPass p) -> Maybe (Located (IdP (GhcPass p)))

splitLHsPatSynTy :: LHsType pass -> ([LHsTyVarBndr pass], LHsContext pass, [LHsTyVarBndr pass], LHsContext pass, LHsType pass)

Decompose a pattern synonym type signature into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsForAllTyInvis :: LHsType pass -> ([LHsTyVarBndr pass], LHsType pass)

Decompose a type of the form forall tvs. body into its constituent parts. Note that only invisible foralls (i.e., forall a., with a dot) are split apart; visible foralls (i.e., forall a ->, with an arrow) are left untouched.

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsQualTy :: LHsType pass -> (LHsContext pass, LHsType pass)

Decompose a type of the form context => body into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (context => ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsSigmaTyInvis :: LHsType pass -> ([LHsTyVarBndr pass], LHsContext pass, LHsType pass)

Decompose a sigma type (of the form forall tvs. context => body) into its constituent parts. Note that only invisible foralls (i.e., forall a., with a dot) are split apart; visible foralls (i.e., forall a ->, with an arrow) are left untouched.

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitHsFunType :: LHsType GhcRn -> ([LHsType GhcRn], LHsType GhcRn)

hsTyGetAppHead_maybe :: LHsType (GhcPass p) -> Maybe (Located (IdP (GhcPass p)))

mkHsOpTy :: LHsType (GhcPass p) -> Located (IdP (GhcPass p)) -> LHsType (GhcPass p) -> HsType (GhcPass p)

mkHsAppTy :: LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

mkHsAppTys :: LHsType (GhcPass p) -> [LHsType (GhcPass p)] -> LHsType (GhcPass p)

mkHsAppKindTy :: XAppKindTy (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

ignoreParens :: LHsType pass -> LHsType pass

hsSigType :: LHsSigType (GhcPass p) -> LHsType (GhcPass p)

hsSigWcType :: LHsSigWcType pass -> LHsType pass

hsLTyVarBndrToType :: LHsTyVarBndr (GhcPass p) -> LHsType (GhcPass p)

Convert a LHsTyVarBndr to an equivalent LHsType.

hsLTyVarBndrsToTypes :: LHsQTyVars (GhcPass p) -> [LHsType (GhcPass p)]

Convert a LHsTyVarBndrs to a list of types. Works on *type* variable only, no kind vars.

hsTyKindSig :: LHsType pass -> Maybe (LHsKind pass)

Get the kind signature of a type, ignoring parentheses:

hsTyKindSig `Maybe ` = Nothing hsTyKindSig `Maybe :: Type -> Type ` = Just `Type -> Type` hsTyKindSig `Maybe :: ((Type -> Type))` = Just `Type -> Type`

This is used to extract the result kind of type synonyms with a CUSK:

type S = (F :: res_kind) ^^^^^^^^

hsConDetailsArgs :: HsConDetails (LHsType a) (Located [LConDeclField a]) -> [LHsType a]

pprHsType :: OutputableBndrId p => HsType (GhcPass p) -> SDoc

pprHsForAll :: OutputableBndrId p => ForallVisFlag -> [LHsTyVarBndr (GhcPass p)] -> LHsContext (GhcPass p) -> SDoc

Prints a forall; When passed an empty list, prints forall ./forall -> only when -dppr-debug is enabled.

pprHsForAllExtra :: OutputableBndrId p => Maybe SrcSpan -> ForallVisFlag -> [LHsTyVarBndr (GhcPass p)] -> LHsContext (GhcPass p) -> SDoc

Version of pprHsForAll that can also print an extra-constraints wildcard, e.g. _ => a -> Bool or (Show a, _) => a -> String. This underscore will be printed when the 'Maybe SrcSpan' argument is a Just containing the location of the extra-constraints wildcard. A special function for this is needed, as the extra-constraints wildcard is removed from the actual context and type, and stored in a separate field, thus just printing the type will not print the extra-constraints wildcard.

pprHsExplicitForAll :: OutputableBndrId p => ForallVisFlag -> Maybe [LHsTyVarBndr (GhcPass p)] -> SDoc

Version of pprHsForAll or pprHsForAllExtra that will always print forall. when passed Just []. Prints nothing if passed Nothing

pprLHsContext :: OutputableBndrId p => LHsContext (GhcPass p) -> SDoc

hsTypeNeedsParens :: PprPrec -> HsType pass -> Bool

hsTypeNeedsParens p t returns True if the type t needs parentheses under precedence p.

parenthesizeHsType :: PprPrec -> LHsType (GhcPass p) -> LHsType (GhcPass p)

parenthesizeHsType p ty checks if hsTypeNeedsParens p ty is true, and if so, surrounds ty with an HsParTy. Otherwise, it simply returns ty.

parenthesizeHsContext :: PprPrec -> LHsContext (GhcPass p) -> LHsContext (GhcPass p)

parenthesizeHsContext p ctxt checks if ctxt is a single constraint c such that hsTypeNeedsParens p c is true, and if so, surrounds c with an HsParTy to form a parenthesized ctxt. Otherwise, it simply returns ctxt unchanged.