GHC.HsToCore.Expr

Documentation

dsExpr :: HsExpr GhcTc -> DsM CoreExpr

dsLExpr :: LHsExpr GhcTc -> DsM CoreExpr

dsLExprNoLP :: LHsExpr GhcTc -> DsM CoreExpr

Variant of dsLExpr that ensures that the result is not levity polymorphic. This should be used when the resulting expression will be an argument to some other function. See Note [Levity polymorphism checking] in GHC.HsToCore.Monad See Note [Levity polymorphism invariants] in GHC.Core

dsLocalBinds :: LHsLocalBinds GhcTc -> CoreExpr -> DsM CoreExpr

dsValBinds :: HsValBinds GhcTc -> CoreExpr -> DsM CoreExpr

dsLit :: HsLit GhcRn -> DsM CoreExpr

dsSyntaxExpr :: SyntaxExpr GhcTc -> [CoreExpr] -> DsM CoreExpr

dsHandleMonadicFailure :: LPat GhcTc -> MatchResult CoreExpr -> FailOperator GhcTc -> DsM CoreExpr