GHC.HsToCore.Match.Literal

Documentation

dsLit :: HsLit GhcRn -> DsM CoreExpr

dsOverLit :: HsOverLit GhcTc -> DsM CoreExpr

Post-typechecker, the HsExpr field of an OverLit contains (an expression for) the literal value itself.

hsLitKey :: Platform -> HsLit GhcTc -> Literal

tidyLitPat :: HsLit GhcTc -> Pat GhcTc

tidyNPat :: HsOverLit GhcTc -> Maybe (SyntaxExpr GhcTc) -> SyntaxExpr GhcTc -> Type -> Pat GhcTc

matchLiterals

Arguments

:: NonEmpty Id

-> Type
Type of the whole case expression

-> NonEmpty (NonEmpty EquationInfo)
All PgLits

-> DsM (MatchResult CoreExpr)

matchNPlusKPats :: NonEmpty Id -> Type -> NonEmpty EquationInfo -> DsM (MatchResult CoreExpr)

matchNPats :: NonEmpty Id -> Type -> NonEmpty EquationInfo -> DsM (MatchResult CoreExpr)

warnAboutIdentities :: DynFlags -> Id -> Type -> DsM ()

warnAboutOverflowedOverLit :: HsOverLit GhcTc -> DsM ()

Emit warnings on overloaded integral literals which overflow the bounds implied by their type.

warnAboutOverflowedLit :: HsLit GhcTc -> DsM ()

Emit warnings on integral literals which overflow the bounds implied by their type.

warnAboutEmptyEnumerations :: FamInstEnvs -> DynFlags -> LHsExpr GhcTc -> Maybe (LHsExpr GhcTc) -> LHsExpr GhcTc -> DsM ()

Warns about [2,3 .. 1] or [b .. a] which return the empty list. For numeric literals, only works for integral types, not floating point.