GHC.HsToCore.Monad

Documentation

type DsM = TcRnIf DsGblEnv DsLclEnv

Desugaring monad. See also TcM.

mapM :: (Traversable t, Monad m) => (a -> m b) -> t a -> m (t b)

mapAndUnzipM :: Applicative m => (a -> m (b, c)) -> [a] -> m ([b], [c])

initDs :: HscEnv -> TcGblEnv -> DsM a -> IO (Messages DsMessage, Maybe a)

Run a DsM action inside the IO monad.

initDsTc :: DsM a -> TcM (Messages DsMessage, Maybe a)

Run a DsM action inside the TcM monad.

initTcDsForSolver :: TcM a -> DsM a

initDsWithModGuts :: HscEnv -> ModGuts -> DsM a -> IO (Messages DsMessage, Maybe a)

Run a DsM action in the context of an existing ModGuts

fixDs :: (a -> DsM a) -> DsM a

foldlM :: (Foldable t, Monad m) => (b -> a -> m b) -> b -> t a -> m b

foldrM :: (Foldable t, Monad m) => (a -> b -> m b) -> b -> t a -> m b

whenGOptM :: GeneralFlag -> TcRnIf gbl lcl () -> TcRnIf gbl lcl ()

unsetGOptM :: GeneralFlag -> TcRnIf gbl lcl a -> TcRnIf gbl lcl a

unsetWOptM :: WarningFlag -> TcRnIf gbl lcl a -> TcRnIf gbl lcl a

xoptM :: Extension -> TcRnIf gbl lcl Bool

class Functor f => Applicative (f :: Type -> Type) where

Minimal complete definition

pure, ((<*>) | liftA2)

Methods

pure :: a -> f a

(<*>) :: f (a -> b) -> f a -> f b

liftA2 :: (a -> b -> c) -> f a -> f b -> f c

(*>) :: f a -> f b -> f b

(<*) :: f a -> f b -> f a

(<$>) :: Functor f => (a -> b) -> f a -> f b

duplicateLocalDs :: Id -> DsM Id

newSysLocalDs :: Scaled Type -> DsM Id

newSysLocalsDs :: [Scaled Type] -> DsM [Id]

newSysLocalMDs :: Type -> DsM Id

newSysLocalsMDs :: [Type] -> DsM [Id]

newFailLocalMDs :: Type -> DsM Id

newUniqueId :: Id -> Mult -> Type -> DsM Id

newPredVarDs :: PredType -> DsM Var

getSrcSpanDs :: DsM SrcSpan

putSrcSpanDs :: SrcSpan -> DsM a -> DsM a

putSrcSpanDsA :: EpAnn ann -> DsM a -> DsM a

mkNamePprCtxDs :: DsM NamePprCtx

newUnique :: TcRnIf gbl lcl Unique

data UniqSupply

Unique Supply

A value of type UniqSupply is unique, and it can supply one distinct Unique. Also, from the supply, one can also manufacture an arbitrary number of further UniqueSupply values, which will be distinct from the first and from all others.

newUniqueSupply :: TcRnIf gbl lcl UniqSupply

getGhcModeDs :: DsM GhcMode

dsGetFamInstEnvs :: DsM FamInstEnvs

dsLookupGlobal :: Name -> DsM TyThing

dsLookupGlobalId :: Name -> DsM Id

dsLookupTyCon :: Name -> DsM TyCon

dsLookupDataCon :: Name -> DsM DataCon

dsLookupConLike :: Name -> DsM ConLike

getCCIndexDsM :: FastString -> DsM CostCentreIndex

See getCCIndexM.

type DsMetaEnv = NameEnv DsMetaVal

data DsMetaVal

Constructors

DsBound Id

DsSplice (HsExpr GhcTc)

dsGetMetaEnv :: DsM (NameEnv DsMetaVal)

dsLookupMetaEnv :: Name -> DsM (Maybe DsMetaVal)

dsExtendMetaEnv :: DsMetaEnv -> DsM a -> DsM a

getPmNablas :: DsM Nablas

Get the current pattern match oracle state. See dsl_nablas.

updPmNablas :: Nablas -> DsM a -> DsM a

Set the pattern match oracle state within the scope of the given action. See dsl_nablas.

addUnspecables :: Set EvId -> DsM a -> DsM a

getUnspecables :: DsM (Set EvId)

dsGetCompleteMatches :: DsM DsCompleteMatches

The COMPLETE pragmas that are in scope.

type DsWarning = (SrcSpan, SDoc)

diagnosticDs :: DsMessage -> DsM ()

Emit a diagnostic for the current source location. In case the diagnostic is a warning, the latter will be ignored and discarded if the relevant WarningFlag is not set in the DynFlags. See Note [Discarding Messages] in Error.

errDsCoreExpr :: DsMessage -> DsM CoreExpr

Issue an error, but return the expression for (), so that we can continue reporting errors.

failWithDs :: DsMessage -> DsM a

failDs :: DsM a

discardWarningsDs :: DsM a -> DsM a

addMessagesDs :: Messages DsMessage -> DsM ()

captureMessagesDs :: DsM a -> DsM (Messages DsMessage, a)

data DsMatchContext

Constructors

DsMatchContext HsMatchContextRn SrcSpan

data EquationInfo

Constructors

EqnMatch

Fields

eqn_pat :: LPat GhcTc

The first pattern of the equation

NB: The location info is used to determine whether the pattern is generated or not. This helps us avoid warnings on patterns that GHC elaborated.

NB: We have already applied decideBangHood to this pattern. See Note [decideBangHood] in GHC.HsToCore.Utils

eqn_rest :: EquationInfo

The rest of the equation after its first pattern

EqnDone (MatchResult CoreExpr)
What to do after match

type EquationInfoNE = EquationInfo

prependPats :: [LPat GhcTc] -> EquationInfo -> EquationInfo

mkEqnInfo :: [LPat GhcTc] -> MatchResult CoreExpr -> EquationInfo

eqnMatchResult :: EquationInfo -> MatchResult CoreExpr

data MatchResult a

This is a value of type a with potentially a CoreExpr-shaped hole in it. This is used to deal with cases where we are potentially handling pattern match failure, and want to later specify how failure is handled.

Constructors

MR_Infallible (DsM a)
We represent the case where there is no hole without a function from CoreExpr, like this, because sometimes we have nothing to put in the hole and so want to be sure there is in fact no hole.

MR_Fallible (CoreExpr -> DsM a)

runMatchResult :: CoreExpr -> MatchResult a -> DsM a

type DsWrapper = CoreExpr -> CoreExpr

idDsWrapper :: DsWrapper

pprRuntimeTrace

Arguments

:: String
header

-> SDoc
information to output

-> CoreExpr
expression

-> DsM CoreExpr

Inject a trace message into the compiled program. Whereas pprTrace prints out information *while compiling*, pprRuntimeTrace captures that information and causes it to be printed *at runtime* using Debug.Trace.trace.

pprRuntimeTrace hdr doc expr

will produce an expression that looks like

trace (hdr + doc) expr

When using this to debug a module that Debug.Trace depends on, it is necessary to import {-# SOURCE #-} Debug.Trace () in that module. We could avoid this inconvenience by wiring in Debug.Trace.trace, but that doesn't seem worth the effort and maintenance cost.

Orphan instances

MonadThings (IOEnv (Env DsGblEnv DsLclEnv))

Instance details

Methods

lookupThing :: Name -> IOEnv (Env DsGblEnv DsLclEnv) TyThing

lookupId :: Name -> IOEnv (Env DsGblEnv DsLclEnv) Id

lookupDataCon :: Name -> IOEnv (Env DsGblEnv DsLclEnv) DataCon

lookupTyCon :: Name -> IOEnv (Env DsGblEnv DsLclEnv) TyCon