GHC.HsToCore.Pmc.Check

Description

Coverage checking step of the Lower Your Guards paper.

Coverage check guard trees (like PmMatch Pre) to get a CheckResult, containing

The set of uncovered values, cr_uncov

And an annotated tree variant (like PmMatch Post) that captures redundancy and inaccessibility information as RedSets annotations

Basically the UA function from Section 5.1, which is an optimised interleaving of U and A from Section 3.2 (Figure 5). The Normalised Refinement Types Nablas are maintained in GHC.HsToCore.Pmc.Solver.

Documentation

newtype CheckAction a

Coverage checking action. Can be composed leftToRight or topToBottom.

Constructors

CA

Fields

unCA :: Nablas -> DsM (CheckResult a)

checkMatchGroup :: PmMatchGroup Pre -> CheckAction (PmMatchGroup Post)

checkGRHSs :: PmGRHSs Pre -> CheckAction (PmGRHSs Post)

checkPatBind :: PmPatBind Pre -> CheckAction (PmPatBind Post)

checkEmptyCase :: PmEmptyCase -> CheckAction PmEmptyCase