GHC.HsToCore.Pmc.Desugar

Description

Desugaring step of the Lower Your Guards paper.

Desugars Haskell source syntax into guard tree variants Pm*. In terms of the paper, this module is concerned with Sections 3.1, Figure 4, in particular.

Documentation

desugarPatBind :: SrcSpan -> Id -> Pat GhcTc -> DsM (PmPatBind Pre)

desugarGRHSs :: SrcSpan -> SDoc -> GRHSs GhcTc (LHsExpr GhcTc) -> DsM (PmGRHSs Pre)

desugarMatches :: [Id] -> NonEmpty (LMatch GhcTc (LHsExpr GhcTc)) -> DsM (PmMatchGroup Pre)

Desugar the non-empty Matches of a MatchGroup.

desugarEmptyCase :: Id -> DsM PmEmptyCase