GHC.HsToCore.Pmc.Ppr

Description

Provides factilities for pretty-printing Nablas in a way appropriate for user facing pattern match warnings.

Documentation

pprUncovered :: Nabla -> [Id] -> SDoc

Pretty-print the guts of an uncovered value vector abstraction, i.e., its components and refutable shapes associated to any mentioned variables.

Example for ([Just p, q], [p :-> [3,4], q :-> [0,5]]):

(Just p) q
    where p is not one of {3, 4}
          q is not one of {0, 5}

When the set of refutable shapes contains more than 3 elements, the additional elements are indicated by "...".