GHC.HsToCore.Pmc.Utils

Description

Utility module for the pattern-match coverage checker.

Documentation

tracePm :: String -> SDoc -> DsM ()

mkPmId :: Type -> DsM Id

Generate a fresh Id of a given type

allPmCheckWarnings :: [WarningFlag]

All warning flags that need to run the pattern match checker.

overlapping :: DynFlags -> HsMatchContext id -> Bool

Check whether the redundancy checker should run (redundancy only)

exhaustive :: DynFlags -> HsMatchContext id -> Bool

Check whether the exhaustiveness checker should run (exhaustiveness only)

redundantBang :: DynFlags -> Bool

Check whether unnecessary bangs should be warned about

exhaustiveWarningFlag :: HsMatchContext id -> Maybe WarningFlag

Denotes whether an exhaustiveness check is supported, and if so, via which WarningFlag it's controlled. Returns Nothing if check is not supported.

isMatchContextPmChecked :: DynFlags -> Origin -> HsMatchContext id -> Bool

Check whether any part of pattern match checking is enabled for this HsMatchContext (does not matter whether it is the redundancy check or the exhaustiveness check).

needToRunPmCheck :: DynFlags -> Origin -> Bool

Return True when any of the pattern match warnings (allPmCheckWarnings) are enabled, in which case we need to run the pattern match checker.