GHC.HsToCore.Types

Description

Various types used during desugaring.

Documentation

type DsM = TcRnIf DsGblEnv DsLclEnv

Desugaring monad. See also TcM.

data DsLclEnv

Local state of the desugarer, extended as we lexically descend

Constructors

DsLclEnv

Fields

dsl_meta :: DsMetaEnv

Template Haskell bindings

dsl_loc :: RealSrcSpan

To put in pattern-matching error msgs

dsl_nablas :: Nablas

See Note [Note [Long-distance information] in GHC.HsToCore.Pmc. The set of reaching values Nablas is augmented as we walk inwards, refined through each pattern match in turn

data DsGblEnv

Global read-only context and state of the desugarer. The statefulness is implemented through IORefs.

Constructors

DsGblEnv

Fields

ds_mod :: Module

ds_fam_inst_env :: FamInstEnv

ds_gbl_rdr_env :: GlobalRdrEnv

ds_unqual :: PrintUnqualified

ds_msgs :: IORef (Messages DecoratedSDoc)

ds_if_env :: (IfGblEnv, IfLclEnv)

ds_complete_matches :: CompleteMatches

ds_cc_st :: IORef CostCentreState

type DsMetaEnv = NameEnv DsMetaVal

data DsMetaVal

Constructors

DsBound Id

DsSplice (HsExpr GhcTc)

type CompleteMatches = [CompleteMatch]