GHC.HsToCore.Usage

Dependency/fingerprinting code (used by GHC.Iface.Make)

mkUsageInfo :: HscEnv -> Module -> ImportedMods -> NameSet -> [FilePath] -> [(Module, Fingerprint)] -> [ModIface] -> IO [Usage]

mkUsedNames :: TcGblEnv -> NameSet

mkDependencies :: UnitId -> [Module] -> TcGblEnv -> IO Dependencies

Extract information from the rename and typecheck phases to produce a dependencies information for the module being compiled.

The second argument is additional dependencies from plugins