GHC.Iface.Ext.Binary

Documentation

readHieFile :: NameCacheUpdater -> FilePath -> IO HieFileResult

Read a HieFile from a FilePath. Can use an existing NameCache.

readHieFileWithVersion :: (HieHeader -> Bool) -> NameCacheUpdater -> FilePath -> IO (Either HieHeader HieFileResult)

Read a HieFile from a FilePath. Can use an existing NameCache. Allows you to specify which versions of hieFile to attempt to read. Left case returns the failing header versions.

type HieHeader = (Integer, ByteString)

writeHieFile :: FilePath -> HieFile -> IO ()

Write a HieFile to the given FilePath, with a proper header and symbol tables for Names and FastStrings

data HieName

Name's get converted into HieName's before being written into .hie files. See toHieName and fromHieName for logic on how to convert between these two types.

Constructors

ExternalName !Module !OccName !SrcSpan

LocalName !OccName !SrcSpan

KnownKeyName !Unique

toHieName :: Name -> HieName

data HieFileResult

Constructors

HieFileResult

Fields

hie_file_result_version :: Integer

hie_file_result_ghc_version :: ByteString

hie_file_result :: HieFile

hieMagic :: [Word8]

The header for HIE files - Capital ASCII letters "HIE".

hieNameOcc :: HieName -> OccName

newtype NameCacheUpdater

A function that atomically updates the name cache given a modifier function. The second result of the modifier function will be the result of the IO action.

Constructors

NCU

Fields

updateNameCache :: forall c. (NameCache -> (NameCache, c)) -> IO c