GHC.Iface.Ext.Types

Documentation

type Span = RealSrcSpan

hieVersion :: Integer

Current version of .hie files

data HieFile

GHC builds up a wealth of information about Haskell source as it compiles it. .hie files are a way of persisting some of this information to disk so that external tools that need to work with haskell source don't need to parse, typecheck, and rename all over again. These files contain:

a simplified AST

nodes are annotated with source positions and types

identifiers are annotated with scope information

the raw bytes of the initial Haskell source

Besides saving compilation cycles, .hie files also offer a more stable interface than the GHC API.

Constructors

HieFile

Fields

hie_hs_file :: FilePath

Initial Haskell source file path

hie_module :: Module

The module this HIE file is for

hie_types :: Array TypeIndex HieTypeFlat

Types referenced in the hie_asts.

See Note [Efficient serialization of redundant type info]

hie_asts :: HieASTs TypeIndex

Type-annotated abstract syntax trees

hie_exports :: [AvailInfo]

The names that this module exports

hie_hs_src :: ByteString

Raw bytes of the initial Haskell source

type TypeIndex = Int

data HieType a

A flattened version of Type.

See Note [Efficient serialization of redundant type info]

Constructors

HTyVarTy Name

HAppTy a (HieArgs a)

HTyConApp IfaceTyCon (HieArgs a)

HForAllTy ((Name, a), ArgFlag) a

HFunTy a a a

HQualTy a a
type with constraint: t1 => t2 (see IfaceDFunTy)

HLitTy IfaceTyLit

HCastTy a

HCoercionTy

type HieTypeFlat = HieType TypeIndex

newtype HieTypeFix

Roughly isomorphic to the original core Type.

Constructors

Roll (HieType HieTypeFix)

newtype HieArgs a

A list of type arguments along with their respective visibilities (ie. is this an argument that would return True for isVisibleArgFlag?).

Constructors

HieArgs [(Bool, a)]

type HiePath = LexicalFastString

pattern HiePath :: FastString -> HiePath

newtype HieASTs a

Mapping from filepaths to the corresponding AST

Constructors

HieASTs

Fields

getAsts :: Map HiePath (HieAST a)

data HieAST a

Constructors

Node

Fields

sourcedNodeInfo :: SourcedNodeInfo a

nodeSpan :: Span

nodeChildren :: [HieAST a]

newtype SourcedNodeInfo a

NodeInfos grouped by source

Constructors

SourcedNodeInfo

Fields

getSourcedNodeInfo :: Map NodeOrigin (NodeInfo a)

data NodeOrigin

Source of node info

Constructors

SourceInfo

GeneratedInfo

data NodeAnnotation

A node annotation

Constructors

NodeAnnotation

Fields

nodeAnnotConstr :: !FastString

name of the AST node constructor

nodeAnnotType :: !FastString

name of the AST node Type

data NodeInfo a

The information stored in one AST node.

The type parameter exists to provide flexibility in representation of types (see Note [Efficient serialization of redundant type info]).

Constructors

NodeInfo

Fields

nodeAnnotations :: Set NodeAnnotation

Annotations

nodeType :: [a]

The Haskell types of this node, if any.

nodeIdentifiers :: NodeIdentifiers a

All the identifiers and their details

pprNodeIdents :: Outputable a => NodeIdentifiers a -> SDoc

pprIdentifier :: Identifier -> SDoc

type Identifier = Either ModuleName Name

type NodeIdentifiers a = Map Identifier (IdentifierDetails a)

data IdentifierDetails a

Information associated with every identifier

We need to include types with identifiers because sometimes multiple identifiers occur in the same span(Overloaded Record Fields and so on)

Constructors

IdentifierDetails

Fields

identType :: Maybe a

identInfo :: Set ContextInfo

data ContextInfo

Different contexts under which identifiers exist

Constructors

Use
regular variable

MatchBind

IEThing IEType
import/export

TyDecl

ValBind
Value binding

Fields

BindType

whether or not the binding is in an instance

Scope

scope over which the value is bound

(Maybe Span)

span of entire binding

PatternBind
Pattern binding

This case is tricky because the bound identifier can be used in two distinct scopes. Consider the following example (with -XViewPatterns)

do (b, a, (a -> True)) <- bar
   foo a

The identifier a has two scopes: in the view pattern (a -> True) and in the rest of the do-block in foo a.

Fields

Scope

scope in the pattern (the variable bound can be used further in the pattern)

Scope

rest of the scope outside the pattern

(Maybe Span)

span of entire binding

ClassTyDecl (Maybe Span)

Decl
Declaration

Fields

DeclType

type of declaration

(Maybe Span)

span of entire binding

TyVarBind Scope TyVarScope
Type variable

RecField RecFieldContext (Maybe Span)
Record field

EvidenceVarBind
Constraint/Dictionary evidence variable binding

Fields

EvVarSource

how did this bind come into being

Scope

scope over which the value is bound

(Maybe Span)

span of the binding site

EvidenceVarUse
Usage of evidence variable

pprBindSpan :: Maybe Span -> SDoc

data EvVarSource

Constructors

EvPatternBind
bound by a pattern match

EvSigBind
bound by a type signature

EvWrapperBind
bound by a hswrapper

EvImplicitBind
bound by an implicit variable

EvInstBind
Bound by some instance of given class

Fields

isSuperInst :: Bool

cls :: Name

EvLetBind EvBindDeps
A direct let binding

newtype EvBindDeps

Eq/Ord instances compare on the converted HieName, as non-exported names may have different uniques after a roundtrip

Constructors

EvBindDeps

Fields

getEvBindDeps :: [Name]

data IEType

Types of imports and exports

Constructors

Import

ImportAs

ImportHiding

Export

data RecFieldContext

Constructors

RecFieldDecl

RecFieldAssign

RecFieldMatch

RecFieldOcc

data BindType

Constructors

RegularBind

InstanceBind

data DeclType

Constructors

FamDec
type or data family

SynDec
type synonym

DataDec
data declaration

ConDec
constructor declaration

PatSynDec
pattern synonym

ClassDec
class declaration

InstDec
instance declaration

data Scope

Constructors

NoScope

LocalScope Span

ModuleScope

data TyVarScope

Scope of a type variable.

This warrants a data type apart from Scope because of complexities introduced by features like -XScopedTypeVariables and -XInstanceSigs. For example, consider:

foo, bar, baz :: forall a. a -> a

Here a is in scope in all the definitions of foo, bar, and baz, so we need a list of scopes to keep track of this. Furthermore, this list cannot be computed until we resolve the binding sites of foo, bar, and baz.

Consequently, a starts with an UnresolvedScope [foo, bar, baz] Nothing which later gets resolved into a ResolvedScopes.

Constructors

ResolvedScopes [Scope]

UnresolvedScope
Unresolved scopes should never show up in the final .hie file

Fields

[Name]

names of the definitions over which the scope spans

(Maybe Span)

the location of the instance/class declaration for the case where the type variable is declared in a method type signature

data HieName

Name's get converted into HieName's before being written into .hie files. See toHieName and fromHieName for logic on how to convert between these two types.

Constructors

ExternalName !Module !OccName !SrcSpan

LocalName !OccName !SrcSpan

KnownKeyName !Unique

hieNameOcc :: HieName -> OccName

toHieName :: Name -> HieName