GHC.Iface.Load

Description

Loading interface files

Documentation

tcLookupImported_maybe :: Name -> TcM (MaybeErr SDoc TyThing)

importDecl :: Name -> IfM lcl (MaybeErr SDoc TyThing)

checkWiredInTyCon :: TyCon -> TcM ()

ifCheckWiredInThing :: TyThing -> IfL ()

loadModuleInterface :: SDoc -> Module -> TcM ModIface

Load interface directly for a fully qualified GenModule. (This is a fairly rare operation, but in particular it is used to load orphan modules in order to pull their instances into the global package table and to handle some operations in GHCi).

loadModuleInterfaces :: SDoc -> [Module] -> TcM ()

Load interfaces for a collection of modules.

loadSrcInterface :: SDoc -> ModuleName -> IsBootInterface -> PkgQual -> RnM ModIface

Load the interface corresponding to an import directive in source code. On a failure, fail in the monad with an error message.

loadSrcInterface_maybe :: SDoc -> ModuleName -> IsBootInterface -> PkgQual -> RnM (MaybeErr SDoc ModIface)

Like loadSrcInterface, but returns a MaybeErr.

loadInterfaceForName :: SDoc -> Name -> TcRn ModIface

Loads the interface for a given Name. Should only be called for an imported name; otherwise loadSysInterface may not find the interface

loadInterfaceForModule :: SDoc -> Module -> TcRn ModIface

Loads the interface for a given Module.

loadInterface :: SDoc -> Module -> WhereFrom -> IfM lcl (MaybeErr SDoc ModIface)

loadSysInterface :: SDoc -> Module -> IfM lcl ModIface

Loads a system interface and throws an exception if it fails

loadUserInterface :: IsBootInterface -> SDoc -> Module -> IfM lcl ModIface

Loads a user interface and throws an exception if it fails. The first parameter indicates whether we should import the boot variant of the module

loadPluginInterface :: SDoc -> Module -> IfM lcl ModIface

findAndReadIface

Arguments

:: HscEnv

-> SDoc
Reason for loading the iface (used for tracing)

-> InstalledModule
The unique identifier of the on-disk module we're looking for

-> Module
The *actual* module we're looking for. We use this to check the consistency of the requirements of the module we read out.

-> IsBootInterface
Looking for .hi-boot or .hi file

-> IO (MaybeErr SDoc (ModIface, FilePath))

readIface :: DynFlags -> NameCache -> Module -> FilePath -> IO (MaybeErr SDoc ModIface)

readIface tries just the one file.

Failed err = file not found, or unreadable, or illegible Succeeded iface = successfully found and parsed

writeIface :: Logger -> Profile -> FilePath -> ModIface -> IO ()

Write interface file

moduleFreeHolesPrecise :: SDoc -> Module -> TcRnIf gbl lcl (MaybeErr SDoc (UniqDSet ModuleName))

Compute the signatures which must be compiled in order to load the interface for a GenModule. The output of this function is always a subset of moduleFreeHoles; it is more precise because in signature p[A=<A>,B=<B>]:B, although the free holes are A and B, B might not depend on A at all!

If this is invoked on a signature, this does NOT include the signature itself; e.g. precise free module holes of p[A=<A>,B=<B>]:B never includes B.

needWiredInHomeIface :: TyThing -> Bool

loadWiredInHomeIface :: Name -> IfM lcl ()

An IfM function to load the home interface for a wired-in thing, so that we're sure that we see its instance declarations and rules See Note [Loading instances for wired-in things]

pprModIfaceSimple :: UnitState -> ModIface -> SDoc

Show a ModIface but don't display details; suitable for ModIfaces stored in the EPT.

ifaceStats :: ExternalPackageState -> SDoc

pprModIface :: UnitState -> ModIface -> SDoc

Show a ModIface

The UnitState is used to pretty-print units

showIface :: Logger -> DynFlags -> UnitState -> NameCache -> FilePath -> IO ()

Read binary interface, and print it out

module GHC.Iface.Errors

Orphan instances

Outputable (Warnings pass)

Instance details

Methods

ppr :: Warnings pass -> SDoc