GHC.Iface.Make

Description

Module for constructing ModIface values (interface files), writing them to disk and comparing two versions to see if recompilation is required.

Documentation

mkPartialIface :: HscEnv -> ModDetails -> ModGuts -> PartialModIface

mkFullIface :: HscEnv -> PartialModIface -> Maybe CgInfos -> IO ModIface

Fully instantiate an interface. Adds fingerprints and potentially code generator produced information.

CgInfos is not available when not generating code (-fno-code), or when not generating interface pragmas (-fomit-interface-pragmas). See also Note [Conveying CAF-info and LFInfo between modules] in GHC.StgToCmm.Types.

mkIfaceTc :: HscEnv -> SafeHaskellMode -> ModDetails -> TcGblEnv -> IO ModIface

Make an interface from the results of typechecking only. Useful for non-optimising compilation, or where we aren't generating any object code at all (HscNothing).

mkIfaceExports :: [AvailInfo] -> [IfaceExport]

coAxiomToIfaceDecl :: CoAxiom br -> IfaceDecl

tyThingToIfaceDecl :: DynFlags -> TyThing -> IfaceDecl