GHC.Iface.Recomp

Description

Module for detecting if recompilation is required

Documentation

checkOldIface :: HscEnv -> ModSummary -> Maybe ModIface -> IO (MaybeValidated ModIface)

Top level function to check if the version of an old interface file is equivalent to the current source file the user asked us to compile. If the same, we can avoid recompilation.

We return on the outside whether the interface file is up to date, providing evidence that is with a ModIface. In the case that it isn't, we may also return a found or provided ModIface. Why we don't always return the old one, if it exists, is unclear to me, except that I tried it and some tests failed (see #18205).

data RecompileRequired

Constructors

UpToDate
everything is up to date, recompilation is not required

NeedsRecompile !CompileReason
Need to compile the module

needsRecompileBecause :: RecompReason -> RecompileRequired

recompThen :: Monad m => m RecompileRequired -> m RecompileRequired -> m RecompileRequired

data MaybeValidated a

Constructors

UpToDateItem a
The item contained is validated to be up to date

OutOfDateItem
The item is are absent altogether or out of date, for the reason given.

Fields

!CompileReason

the reason we need to recompile.

(Maybe a)

The old item, if it exists

outOfDateItemBecause :: RecompReason -> Maybe a -> MaybeValidated a

data RecompReason

Constructors

UnitDepRemoved UnitId

ModulePackageChanged FastString

SourceFileChanged

ThisUnitIdChanged

ImpurePlugin

PluginsChanged

PluginFingerprintChanged

ModuleInstChanged

HieMissing

HieOutdated

SigsMergeChanged

ModuleChanged ModuleName

ModuleRemoved (UnitId, ModuleName)

ModuleAdded (UnitId, ModuleName)

ModuleChangedRaw ModuleName

ModuleChangedIface ModuleName

FileChanged FilePath

CustomReason String

FlagsChanged

OptimFlagsChanged

HpcFlagsChanged

MissingBytecode

MissingObjectFile

MissingDynObjectFile

MissingDynHiFile

MismatchedDynHiFile

ObjectsChanged

LibraryChanged

data CompileReason

Constructors

MustCompile
The .hs file has been touched, or the .o/.hi file does not exist

RecompBecause !RecompReason
The .o/.hi files are up to date, but something else has changed to force recompilation; the String says what (one-line summary)

recompileRequired :: RecompileRequired -> Bool

addFingerprints :: HscEnv -> PartialModIface -> IO ModIface

Add fingerprints for top-level declarations to a ModIface.

See Note [Fingerprinting IfaceDecls]