GHC.Iface.Syntax

Documentation

module GHC.Iface.Type

data IfaceDecl

Constructors

IfaceId

Fields

ifName :: IfaceTopBndr

ifType :: IfaceType

ifIdDetails :: IfaceIdDetails

ifIdInfo :: IfaceIdInfo

IfaceData

Fields

ifName :: IfaceTopBndr

ifBinders :: [IfaceTyConBinder]

ifResKind :: IfaceType

ifCType :: Maybe CType

ifRoles :: [Role]

ifCtxt :: IfaceContext

ifCons :: IfaceConDecls

ifGadtSyntax :: Bool

ifParent :: IfaceTyConParent

IfaceSynonym

Fields

ifName :: IfaceTopBndr

ifRoles :: [Role]

ifBinders :: [IfaceTyConBinder]

ifResKind :: IfaceKind

ifSynRhs :: IfaceType

IfaceFamily

Fields

ifName :: IfaceTopBndr

ifResVar :: Maybe IfLclName

ifBinders :: [IfaceTyConBinder]

ifResKind :: IfaceKind

ifFamFlav :: IfaceFamTyConFlav

ifFamInj :: Injectivity

IfaceClass

Fields

ifName :: IfaceTopBndr

ifRoles :: [Role]

ifBinders :: [IfaceTyConBinder]

ifFDs :: [FunDep IfLclName]

ifBody :: IfaceClassBody

IfaceAxiom

Fields

ifName :: IfaceTopBndr

ifTyCon :: IfaceTyCon

ifRole :: Role

ifAxBranches :: [IfaceAxBranch]

IfacePatSyn

Fields

ifName :: IfaceTopBndr

ifPatIsInfix :: Bool

ifPatMatcher :: (IfExtName, Bool)

ifPatBuilder :: Maybe (IfExtName, Bool)

ifPatUnivBndrs :: [IfaceForAllSpecBndr]

ifPatExBndrs :: [IfaceForAllSpecBndr]

ifPatProvCtxt :: IfaceContext

ifPatReqCtxt :: IfaceContext

ifPatArgs :: [IfaceType]

ifPatTy :: IfaceType

ifFieldLabels :: [FieldLabel]

data IfaceFamTyConFlav

Constructors

IfaceDataFamilyTyCon

IfaceOpenSynFamilyTyCon

IfaceClosedSynFamilyTyCon (Maybe (IfExtName, [IfaceAxBranch]))
Name of associated axiom and branches for pretty printing purposes, or Nothing for an empty closed family without an axiom See Note [Pretty printing via Iface syntax] in GHC.Types.TyThing.Ppr

IfaceAbstractClosedSynFamilyTyCon

IfaceBuiltInSynFamTyCon

data IfaceClassOp

Constructors

IfaceClassOp IfaceTopBndr IfaceType (Maybe (DefMethSpec IfaceType))

data IfaceAT

Constructors

IfaceAT IfaceDecl (Maybe IfaceType)

data IfaceConDecl

Constructors

IfCon

Fields

ifConName :: IfaceTopBndr

ifConWrapper :: Bool

ifConInfix :: Bool

ifConExTCvs :: [IfaceBndr]

ifConUserTvBinders :: [IfaceForAllSpecBndr]

ifConEqSpec :: IfaceEqSpec

ifConCtxt :: IfaceContext

ifConArgTys :: [(IfaceMult, IfaceType)]

ifConFields :: [FieldLabel]

ifConStricts :: [IfaceBang]

ifConSrcStricts :: [IfaceSrcBang]

data IfaceConDecls

Constructors

IfAbstractTyCon

IfDataTyCon [IfaceConDecl]

IfNewTyCon IfaceConDecl

type IfaceEqSpec = [(IfLclName, IfaceType)]

data IfaceExpr

Constructors

IfaceLcl IfLclName

IfaceExt IfExtName

IfaceType IfaceType

IfaceCo IfaceCoercion

IfaceTuple TupleSort [IfaceExpr]

IfaceLam IfaceLamBndr IfaceExpr

IfaceApp IfaceExpr IfaceExpr

IfaceCase IfaceExpr IfLclName [IfaceAlt]

IfaceECase IfaceExpr IfaceType

IfaceLet IfaceBinding IfaceExpr

IfaceCast IfaceExpr IfaceCoercion

IfaceLit Literal

IfaceLitRubbish IfaceType

IfaceFCall ForeignCall IfaceType

IfaceTick IfaceTickish IfaceExpr

data IfaceAlt

Constructors

IfaceAlt IfaceConAlt [IfLclName] IfaceExpr

data IfaceLetBndr

Constructors

IfLetBndr IfLclName IfaceType IfaceIdInfo IfaceJoinInfo

data IfaceJoinInfo

Constructors

IfaceNotJoinPoint

IfaceJoinPoint JoinArity

data IfaceBinding

Constructors

IfaceNonRec IfaceLetBndr IfaceExpr

IfaceRec [(IfaceLetBndr, IfaceExpr)]

data IfaceConAlt

Constructors

IfaceDefault

IfaceDataAlt IfExtName

IfaceLitAlt Literal

type IfaceIdInfo = [IfaceInfoItem]

data IfaceIdDetails

Constructors

IfVanillaId

IfRecSelId (Either IfaceTyCon IfaceDecl) Bool

IfDFunId

data IfaceUnfolding

Constructors

IfCoreUnfold Bool IfaceExpr

IfCompulsory IfaceExpr

IfInlineRule Arity Bool Bool IfaceExpr

IfDFunUnfold [IfaceBndr] [IfaceExpr]

data IfaceInfoItem

Constructors

HsArity Arity

HsStrictness StrictSig

HsCpr CprSig

HsInline InlinePragma

HsUnfold Bool IfaceUnfolding

HsNoCafRefs

HsLevity

HsLFInfo IfaceLFInfo

data IfaceRule

Constructors

IfaceRule

Fields

ifRuleName :: RuleName

ifActivation :: Activation

ifRuleBndrs :: [IfaceBndr]

ifRuleHead :: IfExtName

ifRuleArgs :: [IfaceExpr]

ifRuleRhs :: IfaceExpr

ifRuleAuto :: Bool

ifRuleOrph :: IsOrphan

data IfaceAnnotation

Constructors

IfaceAnnotation

Fields

ifAnnotatedTarget :: IfaceAnnTarget

ifAnnotatedValue :: AnnPayload

type IfaceAnnTarget = AnnTarget OccName

data IfaceClsInst

Constructors

IfaceClsInst

Fields

ifInstCls :: IfExtName

ifInstTys :: [Maybe IfaceTyCon]

ifDFun :: IfExtName

ifOFlag :: OverlapFlag

ifInstOrph :: IsOrphan

data IfaceFamInst

Constructors

IfaceFamInst

Fields

ifFamInstFam :: IfExtName

ifFamInstTys :: [Maybe IfaceTyCon]

ifFamInstAxiom :: IfExtName

ifFamInstOrph :: IsOrphan

data IfaceTickish

Constructors

IfaceHpcTick Module Int

IfaceSCC CostCentre Bool Bool

IfaceSource RealSrcSpan String

data IfaceClassBody

Constructors

IfAbstractClass

IfConcreteClass

Fields

ifClassCtxt :: IfaceContext

ifATs :: [IfaceAT]

ifSigs :: [IfaceClassOp]

ifMinDef :: BooleanFormula IfLclName

data IfaceBang

This corresponds to an HsImplBang; that is, the final implementation decision about the data constructor arg

Constructors

IfNoBang

IfStrict

IfUnpack

IfUnpackCo IfaceCoercion

data IfaceSrcBang

This corresponds to HsSrcBang

Constructors

IfSrcBang SrcUnpackedness SrcStrictness

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{-# UNPACK #-} specified

SrcNoUnpack
{-# NOUNPACK #-} specified

NoSrcUnpack
no unpack pragma

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie ~

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data IfaceAxBranch

Constructors

IfaceAxBranch

Fields

ifaxbTyVars :: [IfaceTvBndr]

ifaxbEtaTyVars :: [IfaceTvBndr]

ifaxbCoVars :: [IfaceIdBndr]

ifaxbLHS :: IfaceAppArgs

ifaxbRoles :: [Role]

ifaxbRHS :: IfaceType

ifaxbIncomps :: [BranchIndex]

data IfaceTyConParent

Constructors

IfNoParent

IfDataInstance IfExtName IfaceTyCon IfaceAppArgs

data IfaceCompleteMatch

Constructors

IfaceCompleteMatch [IfExtName] (Maybe IfaceTyCon)

data IfaceLFInfo

Iface type for LambdaFormInfo. Fields not relevant for imported Ids are omitted in this type.

Constructors

IfLFReEntrant !RepArity

IfLFThunk !Bool !Bool

IfLFCon !Name

IfLFUnknown !Bool

IfLFUnlifted

Binding names

type IfaceTopBndr = Name

A binding top-level Name in an interface file (e.g. the name of an IfaceDecl).

putIfaceTopBndr :: BinHandle -> IfaceTopBndr -> IO ()

getIfaceTopBndr :: BinHandle -> IO IfaceTopBndr

ifaceDeclImplicitBndrs :: IfaceDecl -> [OccName]

visibleIfConDecls :: IfaceConDecls -> [IfaceConDecl]

ifaceDeclFingerprints :: Fingerprint -> IfaceDecl -> [(OccName, Fingerprint)]

freeNamesIfDecl :: IfaceDecl -> NameSet

freeNamesIfRule :: IfaceRule -> NameSet

freeNamesIfFamInst :: IfaceFamInst -> NameSet

pprIfaceExpr :: (SDoc -> SDoc) -> IfaceExpr -> SDoc

Pretty Print an IfaceExpr

The first argument should be a function that adds parens in context that need an atomic value (e.g. function args)

pprIfaceDecl :: ShowSub -> IfaceDecl -> SDoc

newtype AltPpr

Constructors

AltPpr (Maybe (OccName -> SDoc))

data ShowSub

Constructors

ShowSub

Fields

ss_how_much :: ShowHowMuch

ss_forall :: ShowForAllFlag

data ShowHowMuch

Constructors

ShowHeader AltPpr
Header information only, not rhs

ShowSome [OccName] AltPpr
Show only some sub-components. Specifically,

[]

Print all sub-components.

(n:ns)

Print sub-component n with ShowSub = ns; elide other sub-components to ... May 14: the list is max 1 element long at the moment

ShowIface
Everything including GHC-internal information (used in --show-iface)

showToIface :: ShowSub

showToHeader :: ShowSub