GHC.Iface.Type

Documentation

type IfExtName = Name

type IfLclName = FastString

data IfaceType

A kind of universal type, used for types and kinds.

Any time a Type is pretty-printed, it is first converted to an IfaceType before being printed. See Note [Pretty printing via Iface syntax] in GHC.Types.TyThing.Ppr

Constructors

IfaceFreeTyVar TyVar

IfaceTyVar IfLclName

IfaceLitTy IfaceTyLit

IfaceAppTy IfaceType IfaceAppArgs

IfaceFunTy AnonArgFlag IfaceMult IfaceType IfaceType

IfaceForAllTy IfaceForAllBndr IfaceType

IfaceTyConApp IfaceTyCon IfaceAppArgs

IfaceCastTy IfaceType IfaceCoercion

IfaceCoercionTy IfaceCoercion

IfaceTupleTy TupleSort PromotionFlag IfaceAppArgs

type IfacePredType = IfaceType

type IfaceKind = IfaceType

data IfaceCoercion

Constructors

IfaceReflCo IfaceType

IfaceGReflCo Role IfaceType IfaceMCoercion

IfaceFunCo Role IfaceCoercion IfaceCoercion IfaceCoercion

IfaceTyConAppCo Role IfaceTyCon [IfaceCoercion]

IfaceAppCo IfaceCoercion IfaceCoercion

IfaceForAllCo IfaceBndr IfaceCoercion IfaceCoercion

IfaceCoVarCo IfLclName

IfaceAxiomInstCo IfExtName BranchIndex [IfaceCoercion]

IfaceAxiomRuleCo IfLclName [IfaceCoercion]

IfaceUnivCo IfaceUnivCoProv Role IfaceType IfaceType

IfaceSymCo IfaceCoercion

IfaceTransCo IfaceCoercion IfaceCoercion

IfaceNthCo Int IfaceCoercion

IfaceLRCo LeftOrRight IfaceCoercion

IfaceInstCo IfaceCoercion IfaceCoercion

IfaceKindCo IfaceCoercion

IfaceSubCo IfaceCoercion

IfaceFreeCoVar CoVar

IfaceHoleCo CoVar
See Note [Holes in IfaceCoercion]

data IfaceMCoercion

Constructors

IfaceMRefl

IfaceMCo IfaceCoercion

data IfaceUnivCoProv

Constructors

IfacePhantomProv IfaceCoercion

IfaceProofIrrelProv IfaceCoercion

IfacePluginProv String

IfaceCorePrepProv Bool

type IfaceMult = IfaceType

data IfaceTyCon

Constructors

IfaceTyCon

Fields

ifaceTyConName :: IfExtName

ifaceTyConInfo :: IfaceTyConInfo

data IfaceTyConInfo

Constructors

IfaceTyConInfo

Fields

ifaceTyConIsPromoted :: PromotionFlag

ifaceTyConSort :: IfaceTyConSort

mkIfaceTyConInfo :: PromotionFlag -> IfaceTyConSort -> IfaceTyConInfo

data IfaceTyConSort

The various types of TyCons which have special, built-in syntax.

Constructors

IfaceNormalTyCon
a regular tycon

IfaceTupleTyCon !Arity !TupleSort
e.g. (a, b, c) or (#a, b, c#). The arity is the tuple width, not the tycon arity (which is twice the width in the case of unboxed tuples).

IfaceSumTyCon !Arity
e.g. (a | b | c)

IfaceEqualityTyCon
A heterogeneous equality TyCon (i.e. eqPrimTyCon, eqReprPrimTyCon, heqTyCon) that is actually being applied to two types of the same kind. This affects pretty-printing only: see Note [Equality predicates in IfaceType]

data IfaceTyLit

Constructors

IfaceNumTyLit Integer

IfaceStrTyLit FastString

IfaceCharTyLit Char

data IfaceAppArgs

Stores the arguments in a type application as a list. See Note [Suppressing invisible arguments].

Constructors

IA_Nil

IA_Arg IfaceType ArgFlag IfaceAppArgs

type IfaceContext = [IfacePredType]

data IfaceBndr

Constructors

IfaceIdBndr !IfaceIdBndr

IfaceTvBndr !IfaceTvBndr

data IfaceOneShot

Constructors

IfaceNoOneShot

IfaceOneShot

type IfaceLamBndr = (IfaceBndr, IfaceOneShot)

type IfaceTvBndr = (IfLclName, IfaceKind)

type IfaceIdBndr = (IfaceType, IfLclName, IfaceType)

type IfaceTyConBinder = VarBndr IfaceBndr TyConBndrVis

type IfaceForAllSpecBndr = VarBndr IfaceBndr Specificity

type IfaceForAllBndr = VarBndr IfaceBndr ArgFlag

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, TyCoVarBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible Specificity

Required

Bundled Patterns

pattern Specified :: ArgFlag

pattern Inferred :: ArgFlag

data AnonArgFlag

The non-dependent version of ArgFlag. See Note [AnonArgFlag] Appears here partly so that it's together with its friends ArgFlag and ForallVisFlag, but also because it is used in IfaceType, rather early in the compilation chain

Constructors

VisArg
Used for (->): an ordinary non-dependent arrow. The argument is visible in source code.

InvisArg
Used for (=>): a non-dependent predicate arrow. The argument is invisible in source code.

data ShowForAllFlag

Show forall flag

Unconditionally show the forall quantifier with (ShowForAllMust) or when (ShowForAllWhen) the names used are free in the binder or when compiling with -fprint-explicit-foralls.

Constructors

ShowForAllMust

ShowForAllWhen

mkIfaceForAllTvBndr :: ArgFlag -> IfaceTvBndr -> IfaceForAllBndr

Make an IfaceForAllBndr from an IfaceTvBndr.

mkIfaceTyConKind :: [IfaceTyConBinder] -> IfaceKind -> IfaceKind

Build the tyConKind from the binders and the result kind. Keep in sync with mkTyConKind in GHC.Core.TyCon.

ifaceForAllSpecToBndrs :: [IfaceForAllSpecBndr] -> [IfaceForAllBndr]

ifaceForAllSpecToBndr :: IfaceForAllSpecBndr -> IfaceForAllBndr

ifForAllBndrVar :: IfaceForAllBndr -> IfaceBndr

Extract an IfaceBndr from an IfaceForAllBndr.

ifForAllBndrName :: IfaceForAllBndr -> IfLclName

Extract the variable name from an IfaceForAllBndr.

ifaceBndrName :: IfaceBndr -> IfLclName

ifTyConBinderVar :: IfaceTyConBinder -> IfaceBndr

Extract an IfaceBndr from an IfaceTyConBinder.

ifTyConBinderName :: IfaceTyConBinder -> IfLclName

Extract the variable name from an IfaceTyConBinder.

isIfaceLiftedTypeKind :: IfaceKind -> Bool

Given a kind K, is K of the form (TYPE ('BoxedRep 'LiftedRep))?

appArgsIfaceTypes :: IfaceAppArgs -> [IfaceType]

appArgsIfaceTypesArgFlags :: IfaceAppArgs -> [(IfaceType, ArgFlag)]

newtype SuppressBndrSig

Do we want to suppress kind annotations on binders? See Note [Suppressing binder signatures]

Constructors

SuppressBndrSig Bool

newtype UseBndrParens

Constructors

UseBndrParens Bool

newtype PrintExplicitKinds

Constructors

PrintExplicitKinds Bool

pprIfaceType :: IfaceType -> SDoc

pprParendIfaceType :: IfaceType -> SDoc

pprPrecIfaceType :: PprPrec -> IfaceType -> SDoc

pprIfaceContext :: PprPrec -> [IfacePredType] -> SDoc

Prints a context or () if empty You give it the context precedence

pprIfaceContextArr :: [IfacePredType] -> SDoc

Prints "(C a, D b) =>", including the arrow. Used when we want to print a context in a type, so we use funPrec to decide whether to parenthesise a singleton predicate; e.g. Num a => a -> a

pprIfaceIdBndr :: IfaceIdBndr -> SDoc

pprIfaceLamBndr :: IfaceLamBndr -> SDoc

pprIfaceTvBndr :: IfaceTvBndr -> SuppressBndrSig -> UseBndrParens -> SDoc

pprIfaceTyConBinders :: SuppressBndrSig -> [IfaceTyConBinder] -> SDoc

pprIfaceBndrs :: [IfaceBndr] -> SDoc

pprIfaceAppArgs :: IfaceAppArgs -> SDoc

pprParendIfaceAppArgs :: IfaceAppArgs -> SDoc

pprIfaceForAllPart :: [IfaceForAllBndr] -> [IfacePredType] -> SDoc -> SDoc

pprIfaceForAllPartMust :: [IfaceForAllBndr] -> [IfacePredType] -> SDoc -> SDoc

Like pprIfaceForAllPart, but always uses an explicit forall.

pprIfaceForAll :: [IfaceForAllBndr] -> SDoc

Render the "forall ... ." or "forall ... ->" bit of a type.

pprIfaceSigmaType :: ShowForAllFlag -> IfaceType -> SDoc

pprIfaceTyLit :: IfaceTyLit -> SDoc

pprIfaceCoercion :: IfaceCoercion -> SDoc

pprParendIfaceCoercion :: IfaceCoercion -> SDoc

splitIfaceSigmaTy :: IfaceType -> ([IfaceForAllBndr], [IfacePredType], IfaceType)

pprIfaceTypeApp :: PprPrec -> IfaceTyCon -> IfaceAppArgs -> SDoc

pprUserIfaceForAll :: [IfaceForAllBndr] -> SDoc

pprIfaceCoTcApp :: PprPrec -> IfaceTyCon -> [IfaceCoercion] -> SDoc

pprTyTcApp :: PprPrec -> IfaceTyCon -> IfaceAppArgs -> SDoc

pprIfacePrefixApp :: PprPrec -> SDoc -> [SDoc] -> SDoc

ppr_fun_arrow :: IfaceMult -> SDoc

isIfaceTauType :: IfaceType -> Bool

suppressIfaceInvisibles :: PrintExplicitKinds -> [IfaceTyConBinder] -> [a] -> [a]

stripIfaceInvisVars :: PrintExplicitKinds -> [IfaceTyConBinder] -> [IfaceTyConBinder]

stripInvisArgs :: PrintExplicitKinds -> IfaceAppArgs -> IfaceAppArgs

mkIfaceTySubst :: [(IfLclName, IfaceType)] -> IfaceTySubst

substIfaceTyVar :: IfaceTySubst -> IfLclName -> IfaceType

substIfaceAppArgs :: IfaceTySubst -> IfaceAppArgs -> IfaceAppArgs

inDomIfaceTySubst :: IfaceTySubst -> IfaceTvBndr -> Bool

many_ty :: IfaceType