GHC.JS.JStg.Syntax

Description

Domain and Purpose

GHC.JS.JStg.Syntax defines the eDSL that the JS backend's runtime system is written in. Nothing fancy, its just a straightforward deeply embedded DSL.

In general, one should not use these constructors explicitly in the JS backend. Instead, prefer using the combinators in GHC.JS.Make, if those are suitable then prefer using the patterns exported from this module

Deeply embedded JS datatypes

data JStgStat

JavaScript statements, see the ECMA262 Reference for details

Constructors

DeclStat !Ident !(Maybe JStgExpr)
Variable declarations: var foo [= e]

ReturnStat JStgExpr
Return

IfStat JStgExpr JStgStat JStgStat
If

WhileStat Bool JStgExpr JStgStat
While, bool is "do" when True

ForStat JStgStat JStgExpr JStgStat JStgStat
For

ForInStat Bool Ident JStgExpr JStgStat
For-in, bool is "each' when True

SwitchStat JStgExpr [(JStgExpr, JStgStat)] JStgStat
Switch

TryStat JStgStat Ident JStgStat JStgStat
Try

BlockStat [JStgStat]
Blocks

ApplStat JStgExpr [JStgExpr]
Application

UOpStat UOp JStgExpr
Unary operators

AssignStat JStgExpr AOp JStgExpr
Binding form: foo = bar

LabelStat JsLabel JStgStat
Statement Labels, makes me nostalgic for qbasic

BreakStat (Maybe JsLabel)
Break

ContinueStat (Maybe JsLabel)
Continue

FuncStat !Ident [Ident] JStgStat
an explicit function definition

data JStgExpr

JavaScript Expressions

Constructors

ValExpr JVal
All values are trivially expressions

SelExpr JStgExpr Ident
Selection: Obj.foo, see .^

IdxExpr JStgExpr JStgExpr
Indexing: Obj[foo], see .!

InfixExpr Op JStgExpr JStgExpr
Infix Expressions, see JStgExpr pattern synonyms

UOpExpr UOp JStgExpr
Unary Expressions

IfExpr JStgExpr JStgExpr JStgExpr
If-expression

ApplExpr JStgExpr [JStgExpr]
Application

data JVal

JavaScript values

Constructors

JVar Ident
A variable reference

JList [JStgExpr]
A JavaScript list, or what JS calls an Array

JDouble SaneDouble
A Double

JInt Integer
A BigInt

JStr FastString
A String

JRegEx FastString
A Regex

JBool Bool
A Boolean

JHash (UniqMap FastString JStgExpr)
A JS HashMap: {"foo": 0}

JFunc [Ident] JStgStat
A function

data Op

JS Binary Operators. We do not deeply embed the comma operator and the assignment operators

Constructors

EqOp
Equality: ==

StrictEqOp
Strict Equality: ===

NeqOp
InEquality: !=

StrictNeqOp
Strict InEquality !==

GtOp
Greater Than: >

GeOp
Greater Than or Equal: >=

LtOp
Less Than: <

LeOp
Less Than or Equal: <=

AddOp
Addition: +

SubOp
Subtraction: -

MulOp
Multiplication *

DivOp
Division: /

ModOp
Remainder: %

LeftShiftOp
Left Shift: <<

RightShiftOp
Right Shift: >>

ZRightShiftOp
Unsigned RightShift: >>>

BAndOp
Bitwise And: &

BOrOp
Bitwise Or: |

BXorOp
Bitwise XOr: ^

LAndOp
Logical And: &&

LOrOp
Logical Or: ||

InstanceofOp
instanceof

InOp
in

data AOp

JS Unary Operators

Constructors

AssignOp
Vanilla Assignment: =

AddAssignOp
Addition Assignment: +=

SubAssignOp
Subtraction Assignment: -=

data UOp

JS Unary Operators

Constructors

NotOp
Logical Not: !

BNotOp
Bitwise Not: ~

NegOp
Negation: -

PlusOp
Unary Plus: +x

NewOp
new x

TypeofOp
typeof x

DeleteOp
delete x

YieldOp
yield x

VoidOp
void x

PreIncOp
Prefix Increment: ++x

PostIncOp
Postfix Increment: x++

PreDecOp
Prefix Decrement: --x

PostDecOp
Postfix Decrement: x--

type JsLabel = LexicalFastString

A Label used for JStgStat, specifically BreakStat, ContinueStat and of course LabelStat

pattern synonyms over JS operators

pattern New :: JStgExpr -> JStgExpr

pattern synonym for a unary operator new

pattern Not :: JStgExpr -> JStgExpr

pattern synonym for logical not !

pattern Negate :: JStgExpr -> JStgExpr

pattern synonym for unary negation -

pattern Add :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for addition +

pattern Sub :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for subtraction -

pattern Mul :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for multiplication *

pattern Div :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for division *

pattern Mod :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for remainder %

pattern BOr :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for Bitwise Or |

pattern BAnd :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for Bitwise And &

pattern BXor :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for Bitwise XOr ^

pattern BNot :: JStgExpr -> JStgExpr

pattern synonym for Bitwise Not ~

pattern LOr :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for logical Or ||

pattern LAnd :: JStgExpr -> JStgExpr -> JStgExpr

pattern synonym for logical And &&

pattern Int :: Integer -> JStgExpr

pattern synonym to create integer values

pattern String :: FastString -> JStgExpr

pattern synonym to create string values

pattern Var :: Ident -> JStgExpr

pattern synonym to create a local variable reference

pattern PreInc :: JStgExpr -> JStgExpr

pattern synonym for prefix increment ++x

pattern PostInc :: JStgExpr -> JStgExpr

pattern synonym for postfix increment x++

pattern PreDec :: JStgExpr -> JStgExpr

pattern synonym for prefix decrement --x

pattern PostDec :: JStgExpr -> JStgExpr

pattern synonym for postfix decrement --x

Utility

newtype SaneDouble

A newtype wrapper around Double to ensure we never generate a Double that becomes a NaN, see instances for details on sanity.

Constructors

SaneDouble

Fields

unSaneDouble :: Double

pattern Func :: [Ident] -> JStgStat -> JStgExpr

pattern synonym to create an anonymous function

global :: FastString -> JStgExpr

construct a JS reference, intended to refer to a global name

local :: FastString -> JStgExpr

construct a JS reference, intended to refer to a local name