GHC.JS.Make

Description

Domain and Purpose

GHC.JS.Make defines helper functions to ease the creation of JavaScript ASTs as defined in Syntax. Its purpose is twofold: make the EDSL more ergonomic to program in, and make errors in the EDSL look obvious because the EDSL is untyped. It is primarily concerned with injecting terms into the domain of the EDSL to construct JS programs in Haskell.

Strategy

The strategy for this module comes straight from gentzen; where we have two types of helper functions. Functions which inject terms into the EDSL, and combinator functions which operate on terms in the EDSL to construct new terms in the EDSL. Crucially, missing from this module are corresponding elimination or destructing functions which would project information from the EDSL back to Haskell. See Utils for such functions.

Introduction functions

We define various primitive helpers which introduce terms in the EDSL, for example jVar, jLam, and var and jString. Similarly this module exports four typeclasses ToExpr, ToStat, JVarMagic, JSArgument. ToExpr injects values as a JS expression into the EDSL. ToStat injects values as JS statements into the EDSL. JVarMagic provides a polymorphic way to introduce a new name into the EDSL and JSArgument provides a polymorphic way to bind variable names for use in JS functions with different arities.

Combinator functions

The rest of the module defines combinators which create terms in the EDSL from terms in the EDSL. Notable examples are |= and ||=, |= is sugar for AssignStat, it is a binding form that declares foo = bar assuming foo has been already declared. ||= is more sugar on top of |=, it is also a binding form that declares the LHS of |= before calling |= to bind a value, bar, to a variable foo. Other common examples are the if_ and math_ helpers such as math_cos.

Consumers

The entire JS backend consumes this module, e.g., the modules in GHC.StgToJS.*.

Notation

In this module we use ==> in docstrings to show the translation from the JS EDSL domain to JS code. For example, foo ||= bar ==> var foo; foo = bar; should be read as foo ||= bar is in the EDSL domain and results in the JS code var foo; foo = bar; when compiled.

In most cases functions prefixed with a j are monadic because the observably allocate. Notable exceptions are jwhenS, jString and the helpers for HashMaps.

Injection Type classes

The ToJExpr class handles injection of of things into the EDSL as a JS expression

class ToJExpr a where

Things that can be marshalled into javascript values. Instantiate for any necessary data structures.

Minimal complete definition

toJExpr

Methods

toJExpr :: a -> JStgExpr

toJExprFromList :: [a] -> JStgExpr

class ToStat a where

The ToStat class handles injection of of things into the EDSL as a JS statement. This ends up being polymorphic sugar for JS blocks, see helper function expr2stat. Instantiate for any necessary data structures.

Methods

toStat :: a -> JStgStat

class JVarMagic a where

Type class that generates fresh a's for the JS backend. You should almost never need to use this directly. Instead use JSArgument, for examples of how to employ these classes please see jVar, jFunction and call sites in the Rts.

Methods

fresh :: JSM a

class JSArgument args where

Type class that finds the form of arguments required for a JS syntax object. This class gives us a single interface to generate variables for functions that have different arities. Thus with it, we can have only one jFunction which is polymorphic over its arity, instead of jFunction2, jFunction3 and so on.

Methods

argList :: args -> [Ident]

args :: JSM args

Introduction functions

jString :: FastString -> JStgExpr

Convert a FastString to a Javascript String

jLam :: JSArgument args => (args -> JSM JStgStat) -> JSM JStgExpr

Create a new anonymous function. The result is a JExpr expression. Usage:

jLam $ \x -> jVar x + one_
jLam $ \f -> (jLam $ \x -> (f `app` (x `app` x))) `app` (jLam $ \x -> (f `app` (x `app` x)))

jLam' :: JStgStat -> JStgExpr

Special case of jLam where the anonymous function requires no fresh arguments.

jFunction

Arguments

:: JSArgument args

=> Ident
global name

-> (args -> JSM JStgStat)
function body, input is locally unique generated variables

-> JSM JStgStat

Construct a top-level function subject to JS hoisting. This combinator is polymorphic over function arity so you can you use to define a JS syntax object in Haskell, which is a function in JS that takes 2 or 4 or whatever arguments. For a singleton function use the Solo constructor MkSolo. Usage:

an example from the Rts that defines a 1-arity JS function > jFunction (global "h$getReg") ((MkSolo n) -> return $ SwitchStat n getRegCases mempty)

an example of a two argument function from the Rts > jFunction (global "h$bh_lne") ((x, frameSize) -> bhLneStats s x frameSize)

jFunctionSized

Arguments

:: Ident
global name

-> Int
Arity

-> ([JStgExpr] -> JSM JStgStat)
function body, input is locally unique generated variables

-> JSM JStgStat

Construct a top-level function subject to JS hoisting. Special case where the arity cannot be deduced from the args parameter (atleast not without dependent types).

jFunction'

Arguments

:: Ident
global name

-> JSM JStgStat
function body, input is locally unique generated variables

-> JSM JStgStat

Construct a top-level function subject to JS hoisting. Special case where the function binds no parameters

jVar :: (JVarMagic t, ToJExpr t) => (t -> JSM JStgStat) -> JSM JStgStat

Introduce only one new variable into scope for the duration of the enclosed expression. The result is a block statement. Usage:

'jVar $ x -> mconcat [jVar x ||= one_, ...'

jVars :: JSArgument args => (args -> JSM JStgStat) -> JSM JStgStat

Introduce one or many new variables into scope for the duration of the enclosed expression. This function reifies the number of arguments based on the container of the input function. We intentionally avoid lists and instead opt for tuples because lists are not sized in general. The result is a block statement. Usage:

jVars $ (x,y) -> mconcat [ x |= one_,  y |= two_,  x + y]

jFor

Arguments

:: (JStgExpr -> JStgStat)
initialization function

-> (JStgExpr -> JStgExpr)
predicate

-> (JStgExpr -> JStgStat)
step function

-> (JStgExpr -> JStgStat)
body

-> JSM JStgStat

Create a for statement given a function for initialization, a predicate to step to, a step and a body Usage:

 jFor (|= zero_) (.<. Int 65536) preIncrS
        (j -> ...something with the counter j...)

jForIn :: JStgExpr -> (JStgExpr -> JStgStat) -> JSM JStgStat

Create a 'for in' statement. Usage:

jForIn {expression} $ x -> {block involving x}

jForEachIn :: JStgExpr -> (JStgExpr -> JStgStat) -> JSM JStgStat

As with "jForIn" but creating a "for each in" statement.

jTryCatchFinally :: (Ident -> JStgStat) -> (Ident -> JStgStat) -> (Ident -> JStgStat) -> JSM JStgStat

As with "jForIn" but creating a "for each in" statement.

Combinators

Combinators operate on terms in the JS EDSL domain to create new terms in the EDSL domain.

(||=) :: Ident -> JStgExpr -> JStgStat infixl 2

Declare a variable and then Assign the variable to an expression

foo |= expr ==> var foo; foo = expr;

(|=) :: JStgExpr -> JStgExpr -> JStgStat infixl 2

Assign a variable to an expression

foo |= expr ==> var foo = expr;

(.==.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 6

JS infix Equality operators

(.===.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 6

JS infix Equality operators

(.!=.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 6

JS infix Equality operators

(.!==.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 6

JS infix Equality operators

(.!) :: JStgExpr -> JStgExpr -> JStgExpr infixl 8

return the expression at idx of obj

obj .! idx ==> obj[idx]

(.>.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 7

JS infix Ord operators

(.>=.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 7

JS infix Ord operators

(.<.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 7

JS infix Ord operators

(.<=.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 7

JS infix Ord operators

(.<<.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 9

JS infix bit shift operators

(.>>.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 9

JS infix bit shift operators

(.>>>.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 9

JS infix bit shift operators

(.|.) :: JStgExpr -> JStgExpr -> JStgExpr

JS infix bit operators

(.||.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 8

JS infix bit operators

(.&&.) :: JStgExpr -> JStgExpr -> JStgExpr infixl 8

JS infix bit operators

if_ :: JStgExpr -> JStgExpr -> JStgExpr -> JStgExpr

JS if-expression

if_ e1 e2 e3 ==> e1 ? e2 : e3

if10 :: JStgExpr -> JStgExpr

if-expression that returns 1 if condition = true, 0 otherwise

if10 e ==> e ? 1 : 0

if01 :: JStgExpr -> JStgExpr

if-expression that returns 0 if condition = true, 1 otherwise

if01 e ==> e ? 0 : 1

ifS :: JStgExpr -> JStgStat -> JStgStat -> JStgStat

If-expression which returns statements, see related ifBlockS

if e s1 s2 ==> if(e) { s1 } else { s2 }

ifBlockS :: JStgExpr -> [JStgStat] -> [JStgStat] -> JStgStat

If-expression which returns blocks

ifBlockS e s1 s2 ==> if(e) { s1 } else { s2 }

jBlock :: Monoid a => [JSM a] -> JSM a

jIf :: JStgExpr -> JSM JStgStat -> JSM JStgStat -> JSM JStgStat

Version of a JS if-expression which admits monadic actions in its branches

jwhenS :: JStgExpr -> JStgStat -> JStgStat

A when-statement as syntactic sugar via ifS

jwhenS cond block ==> if(cond) { block } else {  }

app :: FastString -> [JStgExpr] -> JStgExpr

an application expression, see related appS

app f xs ==> f(xs)

appS :: FastString -> [JStgExpr] -> JStgStat

A statement application, see the expression form app

returnS :: JStgExpr -> JStgStat

Return a JStgExpr

loop :: JStgExpr -> (JStgExpr -> JStgExpr) -> (JStgExpr -> JSM JStgStat) -> JSM JStgStat

"for" loop with increment at end of body

loopBlockS :: JStgExpr -> (JStgExpr -> JStgExpr) -> (JStgExpr -> [JStgStat]) -> JSM JStgStat

"for" loop with increment at end of body

preIncrS :: JStgExpr -> JStgStat

Prefix-increment a JStgExpr

postIncrS :: JStgExpr -> JStgStat

Postfix-increment a JStgExpr

preDecrS :: JStgExpr -> JStgStat

Prefix-decrement a JStgExpr

postDecrS :: JStgExpr -> JStgStat

Postfix-decrement a JStgExpr

off8 :: JStgExpr -> JStgExpr -> JStgExpr

Byte indexing of o with a 8-bit offset

off16 :: JStgExpr -> JStgExpr -> JStgExpr

Byte indexing of o with a 16-bit offset

off32 :: JStgExpr -> JStgExpr -> JStgExpr

Byte indexing of o with a 32-bit offset

off64 :: JStgExpr -> JStgExpr -> JStgExpr

Byte indexing of o with a 64-bit offset

mask8 :: JStgExpr -> JStgExpr

a bit mask to retrieve the lower 8-bits

mask16 :: JStgExpr -> JStgExpr

a bit mask to retrieve the lower 16-bits

signExtend8 :: JStgExpr -> JStgExpr

Sign-extend/narrow a 8-bit value

signExtend16 :: JStgExpr -> JStgExpr

Sign-extend/narrow a 16-bit value

typeOf :: JStgExpr -> JStgExpr

Given a JStgExpr, return the its type.

returnStack :: JStgStat

assignAllEqual :: HasDebugCallStack => [JStgExpr] -> [JStgExpr] -> JStgStat

assignAll :: [JStgExpr] -> [JStgExpr] -> JStgStat

assignAllReverseOrder :: [JStgExpr] -> [JStgExpr] -> JStgStat

declAssignAll :: [Ident] -> [JStgExpr] -> JStgStat

nullStat :: JStgStat

The empty JS statement

(.^) :: JStgExpr -> FastString -> JStgExpr infixl 8

Select a property prop, from and object obj

obj .^ prop ==> obj.prop

Hash combinators

jhEmpty :: Map k JStgExpr

The empty JS HashMap

jhSingle :: (Ord k, ToJExpr a) => k -> a -> Map k JStgExpr

A singleton JS HashMap

jhAdd :: (Ord k, ToJExpr a) => k -> a -> Map k JStgExpr -> Map k JStgExpr

insert a key-value pair into a JS HashMap

jhFromList :: [(FastString, JStgExpr)] -> JVal

Construct a JS HashMap from a list of key-value pairs

Literals

Literals in the JS EDSL are constants in the Haskell domain. These are useful helper values and never change

null_ :: JStgExpr

The JS literal null

undefined_ :: JStgExpr

The JS literal undefined

false_ :: JStgExpr

The JS literal false

true_ :: JStgExpr

The JS literal true

zero_ :: JStgExpr

The JS literal 0

one_ :: JStgExpr

The JS literal 1

two_ :: JStgExpr

The JS literal 2

three_ :: JStgExpr

The JS literal 3

Math functions

Math functions in the EDSL are literals, with the exception of math_ which is the sole math introduction function.

math_log :: [JStgExpr] -> JStgExpr

math_sin :: [JStgExpr] -> JStgExpr

math_cos :: [JStgExpr] -> JStgExpr

math_tan :: [JStgExpr] -> JStgExpr

math_exp :: [JStgExpr] -> JStgExpr

math_acos :: [JStgExpr] -> JStgExpr

math_asin :: [JStgExpr] -> JStgExpr

math_atan :: [JStgExpr] -> JStgExpr

math_abs :: [JStgExpr] -> JStgExpr

math_pow :: [JStgExpr] -> JStgExpr

math_sqrt :: [JStgExpr] -> JStgExpr

math_asinh :: [JStgExpr] -> JStgExpr

math_acosh :: [JStgExpr] -> JStgExpr

math_atanh :: [JStgExpr] -> JStgExpr

math_cosh :: [JStgExpr] -> JStgExpr

math_sinh :: [JStgExpr] -> JStgExpr

math_tanh :: [JStgExpr] -> JStgExpr

math_expm1 :: [JStgExpr] -> JStgExpr

math_log1p :: [JStgExpr] -> JStgExpr

math_fround :: [JStgExpr] -> JStgExpr

math_min :: [JStgExpr] -> JStgExpr

math_max :: [JStgExpr] -> JStgExpr

Statement helpers

data Solo a

Constructors

MkSolo a

Bundled Patterns

pattern Solo :: a -> Solo a

decl :: Ident -> JStgStat

construct a js declaration with the given identifier

Orphan instances

Num JStgExpr

Instance details

Methods

(+) :: JStgExpr -> JStgExpr -> JStgExpr

(-) :: JStgExpr -> JStgExpr -> JStgExpr

(*) :: JStgExpr -> JStgExpr -> JStgExpr

negate :: JStgExpr -> JStgExpr

abs :: JStgExpr -> JStgExpr

signum :: JStgExpr -> JStgExpr

fromInteger :: Integer -> JStgExpr

Fractional JStgExpr

Instance details

Methods

(/) :: JStgExpr -> JStgExpr -> JStgExpr

recip :: JStgExpr -> JStgExpr

fromRational :: Rational -> JStgExpr