GHC.JS.Syntax

Description

Domain and Purpose

GHC.JS.Syntax defines the Syntax for the JS backend in GHC. It comports with the ECMA-262 although not every production rule of the standard is represented. Code in this module is a fork of JMacro (BSD 3 Clause) by Gershom Bazerman, heavily modified to accomodate GHC's constraints.

Strategy

Nothing fancy in this module, this is a classic deeply embeded AST for JS. We define numerous ADTs and pattern synonyms to make pattern matching and constructing ASTs easier.

Consumers

The entire JS backend consumes this module, e.g., the modules in GHC.StgToJS.*. Please see Make for a module which provides helper functions that use the deeply embedded DSL defined in this module to provide some of the benefits of a shallow embedding.

Deeply embedded JS datatypes

data JStat

JavaScript statements, see the ECMA262 Reference for details

Constructors

DeclStat !Ident !(Maybe JExpr)
Variable declarations: var foo [= e]

ReturnStat JExpr
Return

IfStat JExpr JStat JStat
If

WhileStat Bool JExpr JStat
While, bool is "do" when True

ForStat JStat JExpr JStat JStat
For

ForInStat Bool Ident JExpr JStat
For-in, bool is "each' when True

SwitchStat JExpr [(JExpr, JStat)] JStat
Switch

TryStat JStat Ident JStat JStat
Try

BlockStat [JStat]
Blocks

ApplStat JExpr [JExpr]
Application

UOpStat UOp JExpr
Unary operators

AssignStat JExpr AOp JExpr
Binding form: foo op bar

LabelStat JLabel JStat
Statement Labels, makes me nostalgic for qbasic

BreakStat (Maybe JLabel)
Break

ContinueStat (Maybe JLabel)
Continue

FuncStat !Ident [Ident] JStat
an explicit function definition

data JExpr

JavaScript Expressions

Constructors

ValExpr JVal
All values are trivially expressions

SelExpr JExpr Ident
Selection: Obj.foo, see .^

IdxExpr JExpr JExpr
Indexing: Obj[foo], see .!

InfixExpr Op JExpr JExpr
Infix Expressions, see JExpr pattern synonyms

UOpExpr UOp JExpr
Unary Expressions

IfExpr JExpr JExpr JExpr
If-expression

ApplExpr JExpr [JExpr]
Application

data JVal

JavaScript values

Constructors

JVar Ident
A variable reference

JList [JExpr]
A JavaScript list, or what JS calls an Array

JDouble SaneDouble
A Double

JInt Integer
A BigInt

JStr FastString
A String

JRegEx FastString
A Regex

JHash (UniqMap FastString JExpr)
A JS HashMap: {"foo": 0}

JFunc [Ident] JStat
A function

data Op

JS Binary Operators. We do not deeply embed the comma operator and the assignment operators

Constructors

EqOp
Equality: ==

StrictEqOp
Strict Equality: ===

NeqOp
InEquality: !=

StrictNeqOp
Strict InEquality !==

GtOp
Greater Than: >

GeOp
Greater Than or Equal: >=

LtOp
Less Than: <

LeOp
Less Than or Equal: <=

AddOp
Addition: +

SubOp
Subtraction: -

MulOp
Multiplication *

DivOp
Division: /

ModOp
Remainder: %

LeftShiftOp
Left Shift: <<

RightShiftOp
Right Shift: >>

ZRightShiftOp
Unsigned RightShift: >>>

BAndOp
Bitwise And: &

BOrOp
Bitwise Or: |

BXorOp
Bitwise XOr: ^

LAndOp
Logical And: &&

LOrOp
Logical Or: ||

InstanceofOp
instanceof

InOp
in

data UOp

JS Unary Operators

Constructors

NotOp
Logical Not: !

BNotOp
Bitwise Not: ~

NegOp
Negation: -

PlusOp
Unary Plus: +x

NewOp
new x

TypeofOp
typeof x

DeleteOp
delete x

YieldOp
yield x

VoidOp
void x

PreIncOp
Prefix Increment: ++x

PostIncOp
Postfix Increment: x++

PreDecOp
Prefix Decrement: --x

PostDecOp
Postfix Decrement: x--

data AOp

JS Unary Operators

Constructors

AssignOp
Vanilla Assignment: =

AddAssignOp
Addition Assignment: +=

SubAssignOp
Subtraction Assignment: -=

newtype Ident

A newtype wrapper around FastString for JS identifiers.

Constructors

TxtI

Fields

itxt :: FastString

type JLabel = LexicalFastString

A Label used for JStat, specifically BreakStat, ContinueStat and of course LabelStat

pattern synonyms over JS operators

pattern JNew :: JExpr -> JExpr

pattern synonym for a unary operator new

pattern JNot :: JExpr -> JExpr

pattern synonym for logical not !

pattern JNegate :: JExpr -> JExpr

pattern synonym for unary negation -

pattern JAdd :: JExpr -> JExpr -> JExpr

pattern synonym for addition +

pattern JSub :: JExpr -> JExpr -> JExpr

pattern synonym for subtraction -

pattern JMul :: JExpr -> JExpr -> JExpr

pattern synonym for multiplication *

pattern JDiv :: JExpr -> JExpr -> JExpr

pattern synonym for division *

pattern JMod :: JExpr -> JExpr -> JExpr

pattern synonym for remainder %

pattern JBOr :: JExpr -> JExpr -> JExpr

pattern synonym for Bitwise Or |

pattern JBAnd :: JExpr -> JExpr -> JExpr

pattern synonym for Bitwise And &

pattern JBXor :: JExpr -> JExpr -> JExpr

pattern synonym for Bitwise XOr ^

pattern JBNot :: JExpr -> JExpr

pattern synonym for Bitwise Not ~

pattern JLOr :: JExpr -> JExpr -> JExpr

pattern synonym for logical Or ||

pattern JLAnd :: JExpr -> JExpr -> JExpr

pattern synonym for logical And &&

pattern SatInt :: Integer -> JExpr

pattern synonym to create integer values

pattern JString :: FastString -> JExpr

pattern synonym to create string values

pattern JPreInc :: JExpr -> JExpr

pattern synonym for prefix increment ++x

pattern JPostInc :: JExpr -> JExpr

pattern synonym for postfix increment x++

pattern JPreDec :: JExpr -> JExpr

pattern synonym for prefix decrement --x

pattern JPostDec :: JExpr -> JExpr

pattern synonym for postfix decrement --x

Utility

newtype SaneDouble

A newtype wrapper around Double to ensure we never generate a Double that becomes a NaN, see instances for details on sanity.

Constructors

SaneDouble

Fields

unSaneDouble :: Double

jassignAll :: [JExpr] -> [JExpr] -> JStat

jassignAllEqual :: [JExpr] -> [JExpr] -> JStat

jvar :: FastString -> JExpr