GHC.JS.Transform

Documentation

mapIdent :: (Ident -> JExpr) -> (JExpr -> JExpr, JStat -> JStat)

Map on every variable ident

mapStatIdent :: (Ident -> JExpr) -> JStat -> JStat

mapExprIdent :: (Ident -> JExpr) -> JExpr -> JExpr

identsS :: JStat -> [Ident]

identsV :: JVal -> [Ident]

identsE :: JExpr -> [Ident]

Saturation

jsSaturate :: JMacro a => Maybe FastString -> a -> a

Given an optional prefix, fills in all free variable names with a supply of names generated by the prefix.

Generic traversal (via compos)

class JMacro a where

Compos and ops for generic traversal as defined over the JMacro ADT.

Utility class to coerce the ADT into a regular structure.

Methods

jtoGADT :: a -> JMGadt a

jfromGADT :: JMGadt a -> a

data JMGadt a where

Union type to allow regular traversal by compos.

Constructors

JMGId :: Ident -> JMGadt Ident

JMGStat :: JStat -> JMGadt JStat

JMGExpr :: JExpr -> JMGadt JExpr

JMGVal :: JVal -> JMGadt JVal

class Compos t where

Methods

compos :: (forall a. a -> m a) -> (forall a b. m (a -> b) -> m a -> m b) -> (forall a. t a -> m (t a)) -> t c -> m (t c)

composOp :: Compos t => (forall a. t a -> t a) -> t b -> t b

composOpM :: (Compos t, Monad m) => (forall a. t a -> m (t a)) -> t b -> m (t b)

composOpM_ :: (Compos t, Monad m) => (forall a. t a -> m ()) -> t b -> m ()

composOpFold :: Compos t => b -> (b -> b -> b) -> (forall a. t a -> b) -> t c -> b