GHC.Linker.Deps

Documentation

data LinkDepsOpts

Constructors

LinkDepsOpts

Fields

ldObjSuffix :: !String

Suffix of .o files

ldForceDyn :: !Bool

Always use .dyn_o?

ldOneShotMode :: !Bool

Is the driver in one-shot mode?

ldModuleGraph :: !ModuleGraph

ldUnitEnv :: !UnitEnv

ldPprOpts :: !SDocContext

Rendering options for error messages

ldUseByteCode :: !Bool

Use bytecode rather than objects

ldMsgOpts :: !(DiagnosticOpts IfaceMessage)

Options for diagnostics

ldWays :: !Ways

Enabled ways

ldFinderCache :: !FinderCache

ldFinderOpts :: !FinderOpts

ldLoadIface :: !(SDoc -> Module -> IO (MaybeErr MissingInterfaceError ModIface))

ldLoadByteCode :: !(Module -> IO (Maybe Linkable))

data LinkDeps

Constructors

LinkDeps

Fields

ldNeededLinkables :: [Linkable]

ldAllLinkables :: [Linkable]

ldUnits :: [UnitId]

ldNeededUnits :: UniqDSet UnitId

getLinkDeps :: LinkDepsOpts -> Interp -> LoaderState -> SrcSpan -> [Module] -> IO LinkDeps

Find all the packages and linkables that a set of modules depends on

Return the module and package dependencies for the needed modules. See Note [Object File Dependencies]

Fails with an IO exception if it can't find enough files