GHC.Linker.Static.Utils

Documentation

exeFileName :: ArchOS -> Bool -> Maybe FilePath -> FilePath

Compute the output file name of a program.

StaticLink boolean is used to indicate if the program is actually a static library (e.g., on iOS).

Use the provided filename (if any), otherwise use "main.exe" (Windows), "a.out (otherwise without StaticLink set), "liba.a". In every case, add the extension if it is missing.