GHC.Linker.Unit

Description

Linking Haskell units

Documentation

data UnitLinkOpts

Linker flags collected from units

Constructors

UnitLinkOpts

Fields

hsLibs :: [String]

Haskell libraries (as a list of "-lHSfoo...")

extraLibs :: [String]

External libraries (as a list of "-lfoo...")

otherFlags :: [String]

Extra linker options

collectLinkOpts :: GhcNameVersion -> Ways -> [UnitInfo] -> UnitLinkOpts

collectArchives :: GhcNameVersion -> Ways -> UnitInfo -> IO [FilePath]

getUnitLinkOpts :: GhcNameVersion -> Ways -> UnitEnv -> [UnitId] -> IO UnitLinkOpts

Find all the link options in these and the preload packages, returning (package hs lib options, extra library options, other flags)

getLibs :: GhcNameVersion -> Ways -> UnitEnv -> [UnitId] -> IO [(String, String)]