GHC.Llvm.Ppr

Description

Pretty print LLVM IR Code.

Top level LLVM objects.

ppLlvmModule :: LlvmOpts -> LlvmModule -> SDoc

Print out a whole LLVM module.

ppLlvmComments :: [LMString] -> SDoc

Print out a multi-line comment, can be inside a function or on its own

ppLlvmComment :: LMString -> SDoc

Print out a comment, can be inside a function or on its own

ppLlvmGlobals :: LlvmOpts -> [LMGlobal] -> SDoc

Print out a list of global mutable variable definitions

ppLlvmGlobal :: LlvmOpts -> LMGlobal -> SDoc

Print out a global mutable variable definition

ppLlvmAliases :: [LlvmAlias] -> SDoc

Print out a list of LLVM type aliases.

ppLlvmAlias :: LlvmAlias -> SDoc

Print out an LLVM type alias.

ppLlvmMetas :: LlvmOpts -> [MetaDecl] -> SDoc

Print out a list of LLVM metadata.

ppLlvmMeta :: LlvmOpts -> MetaDecl -> SDoc

Print out an LLVM metadata definition.

ppLlvmFunctionDecls :: LlvmFunctionDecls -> SDoc

Print out a list of function declaration.

ppLlvmFunctionDecl :: LlvmFunctionDecl -> SDoc

Print out a function declaration. Declarations define the function type but don't define the actual body of the function.

ppLlvmFunctions :: LlvmOpts -> LlvmFunctions -> SDoc

Print out a list of function definitions.

ppLlvmFunction :: LlvmOpts -> LlvmFunction -> SDoc

Print out a function definition.

ppVar :: LlvmOpts -> LlvmVar -> SDoc

ppLit :: LlvmOpts -> LlvmLit -> SDoc

Print a literal value. No type.

ppTypeLit :: LlvmOpts -> LlvmLit -> SDoc

ppName :: LlvmOpts -> LlvmVar -> SDoc

Return the variable name or value of the LlvmVar in Llvm IR textual representation (e.g. @x, %y or 42).

ppPlainName :: LlvmOpts -> LlvmVar -> SDoc

Return the variable name or value of the LlvmVar in a plain textual representation (e.g. x, y or 42).