GHC.Parser.Annotation

Documentation

getAnnotation :: ApiAnns -> RealSrcSpan -> AnnKeywordId -> [RealSrcSpan]

Retrieve a list of annotation SrcSpans based on the SrcSpan of the annotated AST element, and the known type of the annotation.

getAndRemoveAnnotation :: ApiAnns -> RealSrcSpan -> AnnKeywordId -> ([RealSrcSpan], ApiAnns)

Retrieve a list of annotation SrcSpans based on the SrcSpan of the annotated AST element, and the known type of the annotation. The list is removed from the annotations.

getAnnotationComments :: ApiAnns -> RealSrcSpan -> [RealLocated AnnotationComment]

Retrieve the comments allocated to the current SrcSpan

Note: A given SrcSpan may appear in multiple AST elements, beware of duplicates

getAndRemoveAnnotationComments :: ApiAnns -> RealSrcSpan -> ([RealLocated AnnotationComment], ApiAnns)

Retrieve the comments allocated to the current SrcSpan, and remove them from the annotations

data ApiAnns

Constructors

ApiAnns

Fields

apiAnnItems :: Map ApiAnnKey [RealSrcSpan]

apiAnnEofPos :: Maybe RealSrcSpan

apiAnnComments :: Map RealSrcSpan [RealLocated AnnotationComment]

apiAnnRogueComments :: [RealLocated AnnotationComment]

type ApiAnnKey = (RealSrcSpan, AnnKeywordId)

data AnnKeywordId

API Annotations exist so that tools can perform source to source conversions of Haskell code. They are used to keep track of the various syntactic keywords that are not captured in the existing AST.

The annotations, together with original source comments are made available in the pm_annotations field of ParsedModule. Comments are only retained if Opt_KeepRawTokenStream is set in DynFlags before parsing.

The wiki page describing this feature is https://gitlab.haskell.org/ghc/ghc/wikis/api-annotations

Note: in general the names of these are taken from the corresponding token, unless otherwise noted See note [Api annotations] above for details of the usage

Constructors

AnnAnyclass

AnnAs

AnnAt

AnnBang
!

AnnBackquote
'`'

AnnBy

AnnCase
case or lambda case

AnnClass

AnnClose
'#)' or '#-}' etc

AnnCloseB
'|)'

AnnCloseBU
'|)', unicode variant

AnnCloseC
'}'

AnnCloseQ
'|]'

AnnCloseQU
'|]', unicode variant

AnnCloseP
')'

AnnCloseS
']'

AnnColon

AnnComma
as a list separator

AnnCommaTuple
in a RdrName for a tuple

AnnDarrow
'=>'

AnnDarrowU
'=>', unicode variant

AnnData

AnnDcolon
'::'

AnnDcolonU
'::', unicode variant

AnnDefault

AnnDeriving

AnnDo

AnnDot
.

AnnDotdot
'..'

AnnElse

AnnEqual

AnnExport

AnnFamily

AnnForall

AnnForallU
Unicode variant

AnnForeign

AnnFunId
for function name in matches where there are multiple equations for the function.

AnnGroup

AnnHeader
for CType

AnnHiding

AnnIf

AnnImport

AnnIn

AnnInfix
'infix' or 'infixl' or 'infixr'

AnnInstance

AnnLam

AnnLarrow
'<-'

AnnLarrowU
'<-', unicode variant

AnnLet

AnnLollyU
The ⊸ unicode arrow

AnnMdo

AnnMinus
-

AnnModule

AnnPercentOne
'%1' -- for HsLinearArrow

AnnNewtype

AnnName
where a name loses its location in the AST, this carries it

AnnOf

AnnOpen
'(#' or '{-# LANGUAGE' etc

AnnOpenB
'(|'

AnnOpenBU
'(|', unicode variant

AnnOpenC
'{'

AnnOpenE
'[e|' or '[e||'

AnnOpenEQ
'[|'

AnnOpenEQU
'[|', unicode variant

AnnOpenP
'('

AnnOpenS
'['

AnnDollar
prefix $ -- TemplateHaskell

AnnDollarDollar
prefix $$ -- TemplateHaskell

AnnPackageName

AnnPattern

AnnPercent
% -- for HsExplicitMult

AnnProc

AnnQualified

AnnRarrow
->

AnnRarrowU
->, unicode variant

AnnRec

AnnRole

AnnSafe

AnnSemi
';'

AnnSimpleQuote
'''

AnnSignature

AnnStatic
static

AnnStock

AnnThen

AnnThTyQuote
double '''

AnnTilde
~

AnnType

AnnUnit
() for types

AnnUsing

AnnVal
e.g. INTEGER

AnnValStr
String value, will need quotes when output

AnnVbar
'|'

AnnVia
via

AnnWhere

Annlarrowtail
-<

AnnlarrowtailU
-<, unicode variant

Annrarrowtail
->

AnnrarrowtailU
->, unicode variant

AnnLarrowtail
-<<

AnnLarrowtailU
-<<, unicode variant

AnnRarrowtail
>>-

AnnRarrowtailU
>>-, unicode variant

data AddAnn

Encapsulated call to addAnnotation, requiring only the SrcSpan of the AST construct the annotation belongs to; together with the AnnKeywordId, this is the key of the annotation map.

This type is useful for places in the parser where it is not yet known what SrcSpan an annotation should be added to. The most common situation is when we are parsing a list: the annotations need to be associated with the AST element that *contains* the list, not the list itself. AddAnn lets us defer adding the annotations until we finish parsing the list and are now parsing the enclosing element; we then apply the AddAnn to associate the annotations. Another common situation is where a common fragment of the AST has been factored out but there is no separate AST node for this fragment (this occurs in class and data declarations). In this case, the annotation belongs to the parent data declaration.

The usual way an AddAnn is created is using the mj ("make jump") function, and then it can be discharged using the ams function.

Constructors

AddAnn AnnKeywordId SrcSpan

mkParensApiAnn :: SrcSpan -> [AddAnn]

Given a SrcSpan that surrounds a HsPar or HsParTy, generate AddAnn values for the opening and closing bordering on the start and end of the span

data AnnotationComment

Constructors

AnnDocCommentNext String
something beginning '-- |'

AnnDocCommentPrev String
something beginning '-- ^'

AnnDocCommentNamed String
something beginning '-- $'

AnnDocSection Int String
a section heading

AnnDocOptions String
doc options (prune, ignore-exports, etc)

AnnLineComment String
comment starting by "--"

AnnBlockComment String
comment in {- -}

data IsUnicodeSyntax

Certain tokens can have alternate representations when unicode syntax is enabled. This flag is attached to those tokens in the lexer so that the original source representation can be reproduced in the corresponding ApiAnnotation

Constructors

UnicodeSyntax

NormalSyntax

unicodeAnn :: AnnKeywordId -> AnnKeywordId

Convert a normal annotation into its unicode equivalent one

data HasE

Some template haskell tokens have two variants, one with an e the other not:

 [| or [e|
 [|| or [e||

This type indicates whether the e is present or not.

Constructors

HasE

NoE

type LRdrName = Located RdrName

AnnKeywordId : AnnOpen, AnnClose,AnnComma, AnnRarrow AnnTilde

May have AnnComma when in a list