GHC.Parser.Lexer.String

Description

This module defines lex states for strings.

This needs to be separate from the normal lexer because the normal lexer automatically includes rules like skipping whitespace or lexing comments, which we don't want in these contexts.

Documentation

data AlexReturn a

Constructors

AlexEOF

AlexError !AlexInput

AlexSkip !AlexInput !Int

AlexToken !AlexInput !Int a

alexScan :: AlexInput -> Int -> AlexReturn a

string_multi_content :: Int