GHC.Parser.Types

Documentation

data SumOrTuple b

Constructors

Sum ConTag Arity (LocatedA b) [EpaLocation] [EpaLocation]
Last two are the locations of the '|' before and after the payload

Tuple [Either (EpAnn EpaLocation) (LocatedA b)]

pprSumOrTuple :: Outputable b => Boxity -> SumOrTuple b -> SDoc

data PatBuilder p

See Note [Ambiguous syntactic categories] and Note [PatBuilder]

Constructors

PatBuilderPat (Pat p)

PatBuilderPar (LocatedA (PatBuilder p)) AnnParen

PatBuilderApp (LocatedA (PatBuilder p)) (LocatedA (PatBuilder p))

PatBuilderAppType (LocatedA (PatBuilder p)) (HsPatSigType GhcPs)

PatBuilderOpApp (LocatedA (PatBuilder p)) (LocatedN RdrName) (LocatedA (PatBuilder p)) (EpAnn [AddEpAnn])

PatBuilderVar (LocatedN RdrName)

PatBuilderOverLit (HsOverLit GhcPs)

data DataConBuilder

An accumulator to build a prefix data constructor, e.g. when parsing MkT A B C, the accumulator will evolve as follows:

 1. PrefixDataConBuilder []        MkT
 2. PrefixDataConBuilder [A]       MkT
 3. PrefixDataConBuilder [A, B]    MkT
 4. PrefixDataConBuilder [A, B, C] MkT
 

There are two reasons we have a separate builder type instead of using HsConDeclDetails GhcPs directly:

It's faster, because OrdList gives us constant-time snoc.

Having a separate type helps ensure that we don't forget to finalize a RecTy into a RecCon (we do that in dataConBuilderDetails).

See Note [PatBuilder] for another builder type used in the parser. Here the technique is similar, but the motivation is different.

Constructors

PrefixDataConBuilder (OrdList (LHsType GhcPs)) (LocatedN RdrName)

InfixDataConBuilder (LHsType GhcPs) (LocatedN RdrName) (LHsType GhcPs)