GHC.Parser.Utils

Documentation

isStmt :: ParserOpts -> String -> Bool

Returns True if passed string is a statement.

hasImport :: ParserOpts -> String -> Bool

Returns True if passed string has an import declaration.

isImport :: ParserOpts -> String -> Bool

Returns True if passed string is an import declaration.

isDecl :: ParserOpts -> String -> Bool

Returns True if passed string is a declaration but not a splice.