GHC.Platform

Description

Platform description

Documentation

data Platform

Platform description

This is used to describe platforms so that we can generate code for them.

Constructors

Platform

Fields

platformArchOS :: !ArchOS

Architecture and OS

platformWordSize :: !PlatformWordSize

Word size

platformByteOrder :: !ByteOrder

Byte order (endianness)

platformUnregisterised :: !Bool

platformHasGnuNonexecStack :: !Bool

platformHasIdentDirective :: !Bool

platformHasSubsectionsViaSymbols :: !Bool

platformIsCrossCompiling :: !Bool

platformLeadingUnderscore :: !Bool

Symbols need underscore prefix

platformTablesNextToCode :: !Bool

Determines whether we will be compiling info tables that reside just before the entry code, or with an indirection to the entry code. See TABLES_NEXT_TO_CODE in includesrtsstorage/InfoTables.h.

platform_constants :: !(Maybe PlatformConstants)

Constants such as structure offsets, type sizes, etc.

data PlatformWordSize

Constructors

PW4
A 32-bit platform

PW8
A 64-bit platform

platformArch :: Platform -> Arch

Platform architecture

platformOS :: Platform -> OS

Platform OS

data ArchOS

Platform architecture and OS.

Constructors

ArchOS

Fields

archOS_arch :: Arch

archOS_OS :: OS

data Arch

Architectures

TODO: It might be nice to extend these constructors with information about what instruction set extensions an architecture might support.

Constructors

ArchUnknown

ArchX86

ArchX86_64

ArchPPC

ArchPPC_64 PPC_64ABI

ArchS390X

ArchSPARC

ArchSPARC64

ArchARM ArmISA [ArmISAExt] ArmABI

ArchAArch64

ArchAlpha

ArchMipseb

ArchMipsel

ArchRISCV64

ArchJavaScript

data OS

Operating systems.

Using OSUnknown to generate code should produce a sensible default, but no promises.

Constructors

OSUnknown

OSLinux

OSDarwin

OSSolaris2

OSMinGW32

OSFreeBSD

OSDragonFly

OSOpenBSD

OSNetBSD

OSKFreeBSD

OSHaiku

OSQNXNTO

OSAIX

OSHurd

data ArmISA

ARM Instruction Set Architecture

Constructors

ARMv5

ARMv6

ARMv7

data ArmISAExt

ARM extensions

Constructors

VFPv2

VFPv3

VFPv3D16

NEON

IWMMX2

data ArmABI

ARM ABI

Constructors

SOFT

SOFTFP

HARD

data PPC_64ABI

PowerPC 64-bit ABI

Constructors

ELF_V1
PowerPC64

ELF_V2
PowerPC64 LE

data ByteOrder

Byte ordering.

Constructors

BigEndian
most-significant-byte occurs in lowest address.

LittleEndian
least-significant-byte occurs in lowest address.

target32Bit :: Platform -> Bool

This predicate tells us whether the platform is 32-bit.

isARM :: Arch -> Bool

osElfTarget :: OS -> Bool

This predicate tells us whether the OS supports ELF-like shared libraries.

osMachOTarget :: OS -> Bool

This predicate tells us whether the OS support Mach-O shared libraries.

osSubsectionsViaSymbols :: OS -> Bool

platformUsesFrameworks :: Platform -> Bool

platformWordSizeInBytes :: Platform -> Int

platformWordSizeInBits :: Platform -> Int

platformMinInt :: Platform -> Integer

Minimum representable Int value for the given platform

platformMaxInt :: Platform -> Integer

Maximum representable Int value for the given platform

platformMaxWord :: Platform -> Integer

Maximum representable Word value for the given platform

platformInIntRange :: Platform -> Integer -> Bool

Test if the given Integer is representable with a platform Int

platformInWordRange :: Platform -> Integer -> Bool

Test if the given Integer is representable with a platform Word

platformCConvNeedsExtension :: Platform -> Bool

For some architectures the C calling convention is that any integer shorter than 64 bits is replaced by its 64 bits representation using sign or zero extension.

data PlatformMisc

Platform-specific settings formerly hard-coded in Config.hs.

These should probably be all be triaged whether they can be computed from other settings or belong in another another place (like Platform above).

Constructors

PlatformMisc

Fields

platformMisc_targetPlatformString :: String

platformMisc_ghcWithInterpreter :: Bool

platformMisc_ghcWithSMP :: Bool

platformMisc_ghcRTSWays :: String

platformMisc_libFFI :: Bool

platformMisc_ghcRtsWithLibdw :: Bool

platformMisc_llvmTarget :: String

data SseVersion

x86 SSE instructions

Constructors

SSE1

SSE2

SSE3

SSE4

SSE42

data BmiVersion

x86 BMI (bit manipulation) instructions

Constructors

BMI1

BMI2

Platform constants

data PlatformConstants

Constructors

PlatformConstants

Fields

pc_CONTROL_GROUP_CONST_291 :: !Int

pc_STD_HDR_SIZE :: !Int

pc_PROF_HDR_SIZE :: !Int

pc_BLOCK_SIZE :: !Int

pc_BLOCKS_PER_MBLOCK :: !Int

pc_TICKY_BIN_COUNT :: !Int

pc_OFFSET_StgRegTable_rR1 :: !Int

pc_OFFSET_StgRegTable_rR2 :: !Int

pc_OFFSET_StgRegTable_rR3 :: !Int

pc_OFFSET_StgRegTable_rR4 :: !Int

pc_OFFSET_StgRegTable_rR5 :: !Int

pc_OFFSET_StgRegTable_rR6 :: !Int

pc_OFFSET_StgRegTable_rR7 :: !Int

pc_OFFSET_StgRegTable_rR8 :: !Int

pc_OFFSET_StgRegTable_rR9 :: !Int

pc_OFFSET_StgRegTable_rR10 :: !Int

pc_OFFSET_StgRegTable_rF1 :: !Int

pc_OFFSET_StgRegTable_rF2 :: !Int

pc_OFFSET_StgRegTable_rF3 :: !Int

pc_OFFSET_StgRegTable_rF4 :: !Int

pc_OFFSET_StgRegTable_rF5 :: !Int

pc_OFFSET_StgRegTable_rF6 :: !Int

pc_OFFSET_StgRegTable_rD1 :: !Int

pc_OFFSET_StgRegTable_rD2 :: !Int

pc_OFFSET_StgRegTable_rD3 :: !Int

pc_OFFSET_StgRegTable_rD4 :: !Int

pc_OFFSET_StgRegTable_rD5 :: !Int

pc_OFFSET_StgRegTable_rD6 :: !Int

pc_OFFSET_StgRegTable_rXMM1 :: !Int

pc_OFFSET_StgRegTable_rXMM2 :: !Int

pc_OFFSET_StgRegTable_rXMM3 :: !Int

pc_OFFSET_StgRegTable_rXMM4 :: !Int

pc_OFFSET_StgRegTable_rXMM5 :: !Int

pc_OFFSET_StgRegTable_rXMM6 :: !Int

pc_OFFSET_StgRegTable_rYMM1 :: !Int

pc_OFFSET_StgRegTable_rYMM2 :: !Int

pc_OFFSET_StgRegTable_rYMM3 :: !Int

pc_OFFSET_StgRegTable_rYMM4 :: !Int

pc_OFFSET_StgRegTable_rYMM5 :: !Int

pc_OFFSET_StgRegTable_rYMM6 :: !Int

pc_OFFSET_StgRegTable_rZMM1 :: !Int

pc_OFFSET_StgRegTable_rZMM2 :: !Int

pc_OFFSET_StgRegTable_rZMM3 :: !Int

pc_OFFSET_StgRegTable_rZMM4 :: !Int

pc_OFFSET_StgRegTable_rZMM5 :: !Int

pc_OFFSET_StgRegTable_rZMM6 :: !Int

pc_OFFSET_StgRegTable_rL1 :: !Int

pc_OFFSET_StgRegTable_rSp :: !Int

pc_OFFSET_StgRegTable_rSpLim :: !Int

pc_OFFSET_StgRegTable_rHp :: !Int

pc_OFFSET_StgRegTable_rHpLim :: !Int

pc_OFFSET_StgRegTable_rCCCS :: !Int

pc_OFFSET_StgRegTable_rCurrentTSO :: !Int

pc_OFFSET_StgRegTable_rCurrentNursery :: !Int

pc_OFFSET_StgRegTable_rHpAlloc :: !Int

pc_OFFSET_stgEagerBlackholeInfo :: !Int

pc_OFFSET_stgGCEnter1 :: !Int

pc_OFFSET_stgGCFun :: !Int

pc_OFFSET_Capability_r :: !Int

pc_OFFSET_bdescr_start :: !Int

pc_OFFSET_bdescr_free :: !Int

pc_OFFSET_bdescr_blocks :: !Int

pc_OFFSET_bdescr_flags :: !Int

pc_SIZEOF_CostCentreStack :: !Int

pc_OFFSET_CostCentreStack_mem_alloc :: !Int

pc_REP_CostCentreStack_mem_alloc :: !Int

pc_OFFSET_CostCentreStack_scc_count :: !Int

pc_REP_CostCentreStack_scc_count :: !Int

pc_OFFSET_StgHeader_ccs :: !Int

pc_OFFSET_StgHeader_ldvw :: !Int

pc_SIZEOF_StgSMPThunkHeader :: !Int

pc_OFFSET_StgEntCounter_allocs :: !Int

pc_REP_StgEntCounter_allocs :: !Int

pc_OFFSET_StgEntCounter_allocd :: !Int

pc_REP_StgEntCounter_allocd :: !Int

pc_OFFSET_StgEntCounter_registeredp :: !Int

pc_OFFSET_StgEntCounter_link :: !Int

pc_OFFSET_StgEntCounter_entry_count :: !Int

pc_SIZEOF_StgUpdateFrame_NoHdr :: !Int

pc_SIZEOF_StgMutArrPtrs_NoHdr :: !Int

pc_OFFSET_StgMutArrPtrs_ptrs :: !Int

pc_OFFSET_StgMutArrPtrs_size :: !Int

pc_SIZEOF_StgSmallMutArrPtrs_NoHdr :: !Int

pc_OFFSET_StgSmallMutArrPtrs_ptrs :: !Int

pc_SIZEOF_StgArrBytes_NoHdr :: !Int

pc_OFFSET_StgArrBytes_bytes :: !Int

pc_OFFSET_StgTSO_alloc_limit :: !Int

pc_OFFSET_StgTSO_cccs :: !Int

pc_OFFSET_StgTSO_stackobj :: !Int

pc_OFFSET_StgStack_sp :: !Int

pc_OFFSET_StgStack_stack :: !Int

pc_OFFSET_StgUpdateFrame_updatee :: !Int

pc_OFFSET_StgFunInfoExtraFwd_arity :: !Int

pc_REP_StgFunInfoExtraFwd_arity :: !Int

pc_SIZEOF_StgFunInfoExtraRev :: !Int

pc_OFFSET_StgFunInfoExtraRev_arity :: !Int

pc_REP_StgFunInfoExtraRev_arity :: !Int

pc_MAX_SPEC_SELECTEE_SIZE :: !Int

pc_MAX_SPEC_AP_SIZE :: !Int

pc_MIN_PAYLOAD_SIZE :: !Int

pc_MIN_INTLIKE :: !Int

pc_MAX_INTLIKE :: !Int

pc_MIN_CHARLIKE :: !Int

pc_MAX_CHARLIKE :: !Int

pc_MUT_ARR_PTRS_CARD_BITS :: !Int

pc_MAX_Vanilla_REG :: !Int

pc_MAX_Float_REG :: !Int

pc_MAX_Double_REG :: !Int

pc_MAX_Long_REG :: !Int

pc_MAX_XMM_REG :: !Int

pc_MAX_Real_Vanilla_REG :: !Int

pc_MAX_Real_Float_REG :: !Int

pc_MAX_Real_Double_REG :: !Int

pc_MAX_Real_XMM_REG :: !Int

pc_MAX_Real_Long_REG :: !Int

pc_RESERVED_C_STACK_BYTES :: !Int

pc_RESERVED_STACK_WORDS :: !Int

pc_AP_STACK_SPLIM :: !Int

pc_WORD_SIZE :: !Int

pc_CINT_SIZE :: !Int

pc_CLONG_SIZE :: !Int

pc_CLONG_LONG_SIZE :: !Int

pc_BITMAP_BITS_SHIFT :: !Int

pc_TAG_BITS :: !Int

pc_LDV_SHIFT :: !Int

pc_ILDV_CREATE_MASK :: !Integer

pc_ILDV_STATE_CREATE :: !Integer

pc_ILDV_STATE_USE :: !Integer

lookupPlatformConstants :: [FilePath] -> IO (Maybe PlatformConstants)

Try to locate "DerivedConstants.h" file in the given dirs and to parse the PlatformConstants from it.

See Note [Platform constants]

platformConstants :: Platform -> PlatformConstants

Shared libraries

platformSOName :: Platform -> FilePath -> FilePath

platformHsSOName :: Platform -> FilePath -> FilePath

platformSOExt :: Platform -> FilePath

genericPlatform :: Platform