GHC.Platform.Reg.Class.Separate

Description

Register classes for architectures which have registers for scalar floating-point values that are separate from all vector registers.

Documentation

data RegClass where

The class of a register. Used in the register allocator. We treat all registers in a class as being interchangeable.

Bundled Patterns

pattern RcInteger :: RegClass

pattern RcFloat :: RegClass

pattern RcVector :: RegClass

pprRegClass :: RegClass -> SDoc

allRegClasses :: [RegClass]