GHC.Platform.Ways

Description

Ways

The central concept of a "way" is that all objects in a given program must be compiled in the same "way". Certain options change parameters of the virtual machine, eg. profiling adds an extra word to the object header, so profiling objects cannot be linked with non-profiling objects.

After parsing the command-line options, we determine which "way" we are building - this might be a combination way, eg. profiling+threaded.

There are two kinds of ways: - RTS only: only affect the runtime system (RTS) and don't affect code generation (e.g. threaded, debug) - Full ways: affect code generation and the RTS (e.g. profiling, dynamic linking)

We then find the "build-tag" associated with this way, and this becomes the suffix used to find .hi files and libraries used in this compilation.

Documentation

data Way

A way

Don't change the constructor order as it us used by waysTag to create a unique tag (e.g. thr_debug_p) which is expected by other tools (e.g. Cabal).

Constructors

WayCustom String
for GHC API clients building custom variants

WayThreaded
(RTS only) Multithreaded runtime system

WayDebug
Debugging, enable trace messages and extra checks

WayProf
Profiling, enable cost-centre stacks and profiling reports

WayTracing
(RTS only) enable event logging (tracing)

WayDyn
Dynamic linking

type Ways = Set Way

hasWay :: Ways -> Way -> Bool

Test if a ways is enabled

addWay :: Way -> Ways -> Ways

Add a way

allowed_combination :: Ways -> Bool

Check if a combination of ways is allowed

wayGeneralFlags :: Platform -> Way -> [GeneralFlag]

Turn these flags on when enabling this way

wayUnsetGeneralFlags :: Platform -> Way -> [GeneralFlag]

Turn these flags off when enabling this way

wayOptc :: Platform -> Way -> [String]

Pass these options to the C compiler when enabling this way

wayOptl :: Platform -> Way -> [String]

Pass these options to linker when enabling this way

wayOptP :: Platform -> Way -> [String]

Pass these options to the preprocessor when enabling this way

wayDesc :: Way -> String

wayRTSOnly :: Way -> Bool

Return true for ways that only impact the RTS, not the generated code

wayTag :: Way -> String

Unique build-tag associated to a way

waysTag :: Ways -> String

Unique tag associated to a list of ways

waysBuildTag :: Ways -> String

Unique build-tag associated to a list of ways

RTS only ways are filtered out because they have no impact on the build.

fullWays :: Ways -> Ways

Filter ways that have an impact on compilation

rtsWays :: Ways -> Ways

Filter RTS-only ways (ways that don't have an impact on compilation)

Host GHC ways

hostWays :: Ways

Host ways.

hostFullWays :: Ways

Host "full" ways (i.e. ways that have an impact on the compilation, not RTS only ways).

These ways must be used when compiling codes targeting the internal interpreter.

hostIsProfiled :: Bool

Consult the RTS to find whether it has been built with profiling enabled.

hostIsDynamic :: Bool

Consult the RTS to find whether GHC itself has been built with dynamic linking. This can't be statically known at compile-time, because we build both the static and dynamic versions together with -dynamic-too.

hostIsThreaded :: Bool

Consult the RTS to find whether it is threaded.

hostIsDebugged :: Bool

Consult the RTS to find whether it is debugged.

hostIsTracing :: Bool

Consult the RTS to find whether it is tracing.