GHC.Runtime.Debugger.Breakpoints

Description

GHC API debugger module for finding and setting breakpoints.

This module is user facing and is at least used by GHCi and `ghc-debugger` to find and set breakpoints.

Documentation

findBreakByLine

Arguments

:: Int
Line number

-> TickArray

-> Maybe (BreakTickIndex, RealSrcSpan)

Find a breakpoint given a Module's TickArray and the line number.

When a line number is specified, the current policy for choosing the best breakpoint is this: - the leftmost complete subexpression on the specified line, or - the leftmost subexpression starting on the specified line, or - the rightmost subexpression enclosing the specified line

findBreakByCoord :: (Int, Int) -> TickArray -> Maybe (BreakTickIndex, RealSrcSpan)

Find a breakpoint in the TickArray of a module, given a line number and a column coordinate.

leftmostLargestRealSrcSpan :: RealSrcSpan -> RealSrcSpan -> Ordering

enclosingTickSpan :: TickArray -> SrcSpan -> RealSrcSpan

Returns the span of the largest tick containing the srcspan given

resolveFunctionBreakpoint :: GhcMonad m => String -> m (Either SDoc (Module, ModuleInfo, String))

Process and validate the user string of form [Module.]function into the relevant module information and function name.

Validation guarantees 1. The module exists 2. The identifier is in an interpreted module 3. The identifier has a breakpoint entry in the module's ModBreaks

Returns either an error SDoc or the Module and ModuleInfo for the relevant module paired with the function name

See also Note [Setting Breakpoints by Id]

findBreakForBind

Arguments

:: String
Name of bind to break at

-> ModBreaks

-> [(BreakTickIndex, RealSrcSpan)]

The aim of this function is to find the breakpoints for all the RHSs of the equations corresponding to a binding. So we find all breakpoints for (a) this binder only (it maybe a top-level or a nested declaration) (b) that do not have an enclosing breakpoint

type TickArray = Array Int [(BreakTickIndex, RealSrcSpan)]

Maps line numbers to the breakpoint ticks existing at that line for a module.

makeModuleLineMap :: GhcMonad m => Module -> m (Maybe TickArray)

Construct the TickArray for the given module.

getModBreak :: GhcMonad m => Module -> m (Maybe ModBreaks)

Get the ModBreaks of the given Module when available

type BreakpointOccurrences = ModuleEnv (IntMap [InternalBreakpointId])

A source-level breakpoint may have been inlined into many occurrences, now referred by InternalBreakpointId. When a breakpoint is set on a certain source breakpoint, it means all *ocurrences* of that breakpoint across modules should be stopped at -- hence we keep a trie from BreakpointId to the list of internal break ids using it. See also Note [Breakpoint identifiers]

lookupBreakpointOccurrences :: BreakpointOccurrences -> BreakpointId -> Maybe [InternalBreakpointId]

Lookup all InternalBreakpointIds matching the given BreakpointId Nothing if BreakpointId not in map

mkBreakpointOccurrences :: GhcMonad m => m BreakpointOccurrences

Construct a mapping from Source BreakpointIds to InternalBreakpointIds from the given list of ModInfos

getCurrentBreakSpan :: GhcMonad m => m (Maybe SrcSpan)

getCurrentBreakModule :: GhcMonad m => m (Maybe Module)