GHC.Runtime.Interpreter.JS

Description

JavaScript interpreter

See Note [The JS interpreter]

Documentation

spawnJSInterp :: JSInterpConfig -> IO (ExtInterpInstance JSInterpExtra)

Spawn a JS interpreter

Run NodeJS with "ghc-interp.js" loaded in. Then load GHCi.Server and its deps (including the rts) and run GHCi.Server.defaultServer.

jsLinkRts :: Logger -> TmpFs -> TempDir -> StgToJSConfig -> UnitEnv -> ExtInterpInstance JSInterpExtra -> IO ()

Link JS RTS

jsLinkInterp :: Logger -> TmpFs -> TempDir -> StgToJSConfig -> UnitEnv -> ExtInterpInstance JSInterpExtra -> IO ()

Link JS interpreter

jsLinkObject :: Logger -> TmpFs -> TempDir -> StgToJSConfig -> UnitEnv -> ExtInterpInstance JSInterpExtra -> FilePath -> [ExportedFun] -> IO ()

Link an object file using the given functions as roots

jsLinkObjects :: Logger -> TmpFs -> TempDir -> StgToJSConfig -> UnitEnv -> ExtInterpInstance JSInterpExtra -> [FilePath] -> (ExportedFun -> Bool) -> IO ()

Link object files

jsLoadFile :: ExtInterpInstance JSInterpExtra -> FilePath -> IO ()

Load a JS file in the interpreter

jsRunServer :: ExtInterpInstance JSInterpExtra -> IO ()

Run JS server

Reexported for convenience

mkExportedModFuns :: Module -> [FastString] -> [ExportedFun]

Given a Module and a set of symbols in the module, package these into an ExportedFun.