GHC.Stg.InferTags.Types

Documentation

type InferStgTopBinding = GenStgTopBinding 'InferTaggedBinders

type InferStgBinding = GenStgBinding 'InferTaggedBinders

type InferStgExpr = GenStgExpr 'InferTaggedBinders

type InferStgRhs = GenStgRhs 'InferTaggedBinders

type InferStgAlt = GenStgAlt 'InferTaggedBinders

type TagSigEnv = IdEnv TagSig

data TagEnv p

Constructors

TE

Fields

te_env :: TagSigEnv

te_get :: BinderP p -> Id

te_bytecode :: !Bool

noSig :: TagEnv p -> BinderP p -> (Id, TagSig)

extendSigEnv :: TagEnv p -> [(Id, TagSig)] -> TagEnv p

initEnv :: Bool -> TagEnv 'CodeGen

combineAltInfo :: TagInfo -> TagInfo -> TagInfo

getBinderId :: TagEnv p -> BinderP p -> Id

makeTagged :: TagEnv p -> TagEnv 'InferTaggedBinders

Simple convert env to a env of the 'InferTaggedBinders pass with no other changes.

lookupSig :: TagEnv p -> Id -> Maybe TagSig

Look up a sig in the given env

lookupInfo :: TagEnv p -> StgArg -> TagInfo

Look up a sig in the env or derive it from information in the arg itself.

isDunnoSig :: TagSig -> Bool

isTaggedInfo :: TagInfo -> Bool

module GHC.Stg.InferTags.TagSig