GHC.Stg.Lift.Config

Description

Configuration options for Lift the lambda lifter.

Documentation

data StgLiftConfig

Constructors

StgLiftConfig

Fields

c_targetProfile :: !Profile

c_liftLamsRecArgs :: !(Maybe Int)

Maximum number of arguments after lambda lifting a recursive function.

c_liftLamsNonRecArgs :: !(Maybe Int)

Maximum number of arguments after lambda lifting non-recursive function.

c_liftLamsKnown :: !Bool

Lambda lift even when this turns a known call into an unknown call.