GHC.Stg.Lift.Types

Documentation

data Skeleton

Captures details of the syntax tree relevant to the cost model, such as closures, multi-shot lambdas and case expressions.

Constructors

ClosureSk

Fields

!Id

!DIdSet

free vars

!Skeleton

RhsSk

Fields

!Card

how often the RHS was entered

!Skeleton

AltSk !Skeleton !Skeleton

BothSk !Skeleton !Skeleton

NilSk

bothSk :: Skeleton -> Skeleton -> Skeleton

altSk :: Skeleton -> Skeleton -> Skeleton

rhsSk :: Card -> Skeleton -> Skeleton

data BinderInfo

The type used in binder positions in GenStgExprs.

Constructors

BindsClosure !Id !Bool
Let(-no-escape)-bound thing with a flag indicating whether it occurs as an argument or in a nullary application (see GHC.Stg.Lift.Analysis).

BoringBinder !Id
Every other kind of binder

binderInfoBndr :: BinderInfo -> Id

Gets the bound Id out a BinderInfo.

binderInfoOccursAsArg :: BinderInfo -> Maybe Bool

Returns Nothing for BoringBinders and Just the flag indicating occurrences as argument or in a nullary applications otherwise.