GHC.Stg.Pipeline

Documentation

data StgPipelineOpts

Constructors

StgPipelineOpts

Fields

stgPipeline_phases :: ![StgToDo]

Spec of what stg-to-stg passes to do

stgPipeline_lint :: !(Maybe DiagOpts)

Should we lint the STG at various stages of the pipeline?

stgPipeline_pprOpts :: !StgPprOpts

stgPlatform :: !Platform

stgPipeline_forBytecode :: !Bool

data StgToDo

Optional Stg-to-Stg passes.

Constructors

StgCSE
Common subexpression elimination

StgLiftLams StgLiftConfig
Lambda lifting closure variables, trading stack/register allocation for heap allocation

StgStats

StgUnarise
Mandatory unarise pass, desugaring unboxed tuple and sum binders

StgBcPrep
Mandatory when compiling to bytecode

StgDoNothing
Useful for building up getStgToDo

stg2stg

Arguments

:: Logger

-> [Var]
extra vars in scope from GHCi

-> StgPipelineOpts

-> Module
module being compiled

-> [StgTopBinding]
input program

-> IO ([CgStgTopBinding], StgCgInfos)

type StgCgInfos = NameEnv TagSig

Information to be exposed in interface files which is produced by the stg2stg passes.