GHC.Stg.Utils

Documentation

mkStgAltTypeFromStgAlts :: forall (p :: StgPass). Id -> [GenStgAlt p] -> AltType

bindersOf :: forall (a :: StgPass). BinderP a ~ Id => GenStgBinding a -> [Id]

bindersOfX :: forall (a :: StgPass). GenStgBinding a -> [BinderP a]

bindersOfTop :: forall (a :: StgPass). BinderP a ~ Id => GenStgTopBinding a -> [Id]

bindersOfTopBinds :: forall (a :: StgPass). BinderP a ~ Id => [GenStgTopBinding a] -> [Id]

stripStgTicksTop :: forall (p :: StgPass). (StgTickish -> Bool) -> GenStgExpr p -> ([StgTickish], GenStgExpr p)

Strip ticks of a given type from an STG expression.

stripStgTicksTopE :: forall (p :: StgPass). (StgTickish -> Bool) -> GenStgExpr p -> GenStgExpr p

Strip ticks of a given type from an STG expression returning only the expression.

idArgs :: [StgArg] -> [Id]

mkUnarisedId :: MonadUnique m => FastString -> NvUnaryType -> m Id

mkUnarisedIds :: MonadUnique m => FastString -> [NvUnaryType] -> m [Id]

allowTopLevelConApp :: Platform -> Bool -> Module -> DataCon -> [StgArg] -> Bool

Do we allow the given top-level (static) ConApp?