GHC.StgToCmm.Closure

Documentation

type DynTag = Int

tagForCon :: Platform -> DataCon -> DynTag

isSmallFamily :: Platform -> Int -> Bool

idPrimRep :: Id -> PrimRep

Assumes that there is precisely one PrimRep of the type. This assumption holds after unarise. See Note [Post-unarisation invariants]

isVoidRep :: PrimRep -> Bool

isGcPtrRep :: PrimRep -> Bool

addIdReps :: [NonVoid Id] -> [NonVoid (PrimRep, Id)]

Assumes that Ids have one PrimRep, which holds after unarisation. See Note [Post-unarisation invariants]

addArgReps :: [NonVoid StgArg] -> [NonVoid (PrimRep, StgArg)]

Assumes that arguments have one PrimRep, which holds after unarisation. See Note [Post-unarisation invariants]

argPrimRep :: StgArg -> PrimRep

Assumes that the argument has one PrimRep, which holds after unarisation. See Note [Post-unarisation invariants]

newtype NonVoid a

Constructors

NonVoid a

fromNonVoid :: NonVoid a -> a

nonVoidIds :: [Id] -> [NonVoid Id]

nonVoidStgArgs :: [StgArg] -> [NonVoid StgArg]

assertNonVoidIds :: [Id] -> [NonVoid Id]

Used in places where some invariant ensures that all these Ids are non-void; e.g. constructor field binders in case expressions. See Note [Post-unarisation invariants] in GHC.Stg.Unarise.

assertNonVoidStgArgs :: [StgArg] -> [NonVoid StgArg]

Used in places where some invariant ensures that all these arguments are non-void; e.g. constructor arguments. See Note [Post-unarisation invariants] in GHC.Stg.Unarise.

LambdaFormInfo

data LambdaFormInfo

Information about an identifier, from the code generator's point of view. Every identifier is bound to a LambdaFormInfo in the environment, which gives the code generator enough info to be able to tail call or return that identifier.

data StandardFormInfo

StandardFormInfo tells whether this thunk has one of a small number of standard forms

mkLFThunk :: Type -> TopLevelFlag -> [Id] -> UpdateFlag -> LambdaFormInfo

mkLFReEntrant :: TopLevelFlag -> [Id] -> [Id] -> ArgDescr -> LambdaFormInfo

mkConLFInfo :: DataCon -> LambdaFormInfo

mkSelectorLFInfo :: Id -> Int -> Bool -> LambdaFormInfo

mkApLFInfo :: Id -> UpdateFlag -> Arity -> LambdaFormInfo

mkLFImported :: Id -> LambdaFormInfo

mkLFArgument :: Id -> LambdaFormInfo

mkLFLetNoEscape :: LambdaFormInfo

mkLFStringLit :: LambdaFormInfo

lfDynTag :: Platform -> LambdaFormInfo -> DynTag

Return the tag in the low order bits of a variable bound to this LambdaForm

isLFThunk :: LambdaFormInfo -> Bool

isLFReEntrant :: LambdaFormInfo -> Bool

lfUpdatable :: LambdaFormInfo -> Bool

Used by other modules

data CgLoc

Constructors

CmmLoc CmmExpr

LneLoc BlockId [LocalReg]

type SelfLoopInfo = (Id, BlockId, [LocalReg])

data CallMethod

Constructors

EnterIt

JumpToIt BlockId [LocalReg]

ReturnIt

SlowCall

DirectEntry CLabel RepArity

nodeMustPointToIt :: Profile -> LambdaFormInfo -> Bool

isKnownFun :: LambdaFormInfo -> Bool

funTag :: Platform -> ClosureInfo -> DynTag

tagForArity :: Platform -> RepArity -> DynTag

data CallOpts

Constructors

CallOpts

Fields

co_profile :: !Profile

Platform profile

co_loopification :: !Bool

Loopification enabled (cf -floopification)

co_ticky :: !Bool

Ticky profiling enabled (cf -ticky)

getCallMethod :: CallOpts -> Name -> Id -> LambdaFormInfo -> RepArity -> RepArity -> CgLoc -> Maybe SelfLoopInfo -> CallMethod

ClosureInfo

data ClosureInfo

mkClosureInfo :: Profile -> Bool -> Id -> LambdaFormInfo -> Int -> Int -> String -> ClosureInfo

mkCmmInfo :: ClosureInfo -> Id -> CostCentreStack -> CmmInfoTable

Convert from ClosureInfo to CmmInfoTable.

Inspection

closureLFInfo :: ClosureInfo -> LambdaFormInfo

closureName :: ClosureInfo -> Name

Labels

closureInfoLabel :: ClosureInfo -> CLabel

staticClosureLabel :: Platform -> ClosureInfo -> CLabel

closureSlowEntryLabel :: Platform -> ClosureInfo -> CLabel

closureLocalEntryLabel :: Platform -> ClosureInfo -> CLabel

Predicates

closureUpdReqd :: ClosureInfo -> Bool

closureReEntrant :: ClosureInfo -> Bool

closureFunInfo :: ClosureInfo -> Maybe (RepArity, ArgDescr)

isToplevClosure :: ClosureInfo -> Bool

blackHoleOnEntry :: ClosureInfo -> Bool

isStaticClosure :: ClosureInfo -> Bool

InfoTables

mkDataConInfoTable :: Profile -> DataCon -> ConInfoTableLocation -> Bool -> Int -> Int -> CmmInfoTable

cafBlackHoleInfoTable :: CmmInfoTable

indStaticInfoTable :: CmmInfoTable

staticClosureNeedsLink :: Bool -> CmmInfoTable -> Bool