GHC.StgToCmm.ExtCode

Description

Our extended FCode monad.

Documentation

data CmmParse a

Does a computation in the FCode monad, with a current environment and a list of local declarations. Returns the resulting list of declarations.

unEC :: CmmParse a -> String -> Env -> Decls -> FCode (Decls, a)

data Named

The environment contains variable definitions or blockids.

Constructors

VarN CmmExpr
Holds CmmLit(CmmLabel ..) which gives the label type, eg, RtsLabel, ForeignLabel, CmmLabel etc.

FunN UnitId
A function name from this unit

LabelN BlockId
A blockid of some code or data.

type Env = UniqFM FastString Named

An environment of named things.

loopDecls :: CmmParse a -> CmmParse a

Takes the variable declarations and imports from the monad and makes an environment, which is looped back into the computation. In this way, we can have embedded declarations that scope over the whole procedure, and imports that scope over the entire module. Discards the local declaration contained within decl'

getEnv :: CmmParse Env

Get the current environment from the monad.

withName :: String -> CmmParse a -> CmmParse a

Set context name for a sub-parse

getName :: CmmParse String

Get the current context name from the monad

newLocal

Arguments

:: CmmType
data type

-> FastString
name of variable

-> CmmParse LocalReg
register holding the value

Create a fresh local variable of a given type.

newLabel :: FastString -> CmmParse BlockId

Allocate a fresh label.

newBlockId :: MonadUnique m => m BlockId

newFunctionName

Arguments

:: FastString
name of the function

-> UnitId
package of the current module

-> ExtCode

Add a local function to the environment.

newImport :: (FastString, CLabel) -> CmmParse ()

Add an imported foreign label to the list of local declarations. If this is done at the start of the module the declaration will scope over the whole module.

lookupLabel :: FastString -> CmmParse BlockId

Lookup the BlockId bound to the label with this name. If one hasn't been bound yet, create a fresh one based on the Unique of the name.

lookupName :: FastString -> CmmParse CmmExpr

Lookup the location of a named variable. Unknown names are treated as if they had been 'import'ed from the runtime system. This saves us a lot of bother in the RTS sources, at the expense of deferring some errors to link time.

code :: FCode a -> CmmParse a

Lift an FCode computation into the CmmParse monad

emit :: CmmAGraph -> CmmParse ()

emitLabel :: BlockId -> CmmParse ()

emitAssign :: CmmReg -> CmmExpr -> CmmParse ()

emitStore :: CmmExpr -> CmmExpr -> CmmParse ()

getCode :: CmmParse a -> CmmParse CmmAGraph

getCodeR :: CmmParse a -> CmmParse (a, CmmAGraph)

getCodeScoped :: CmmParse a -> CmmParse (a, CmmAGraphScoped)

emitOutOfLine :: BlockId -> CmmAGraphScoped -> CmmParse ()

withUpdFrameOff :: UpdFrameOffset -> CmmParse () -> CmmParse ()

getUpdFrameOff :: CmmParse UpdFrameOffset

getProfile :: CmmParse Profile

getPlatform :: CmmParse Platform

getPtrOpts :: CmmParse PtrOpts