GHC.StgToCmm.InfoTableProv

Documentation

emitIpeBufferListNode

Arguments

:: Module

-> [InfoProvEnt]

-> DUniqSupply
Symbols created source uniques deterministically All uniques must be created from this supply. NB: If you are creating a new symbol within this function, make sure it is local only (as in not externallyVisibleCLabel). If you need it to be global, reconsider the comment on the call of emitIpeBufferListNode in Cmm.Parser.

-> FCode DUniqSupply