GHC.StgToCmm.Monad

Documentation

data FCode a

initC :: IO CgState

initFCodeState :: Platform -> FCodeState

runC :: StgToCmmConfig -> FCodeState -> CgState -> FCode a -> (a, CgState)

fixC :: (a -> FCode a) -> FCode a

newUnique :: FCode Unique

emitLabel :: BlockId -> FCode ()

emit :: CmmAGraph -> FCode ()

emitDecl :: CmmDecl -> FCode ()

emitProcWithConvention :: Convention -> Maybe CmmInfoTable -> CLabel -> [CmmFormal] -> CmmAGraphScoped -> FCode ()

emitProcWithStackFrame :: Convention -> Maybe CmmInfoTable -> CLabel -> [CmmFormal] -> [CmmFormal] -> CmmAGraphScoped -> Bool -> FCode ()

emitOutOfLine :: BlockId -> CmmAGraphScoped -> FCode ()

emitAssign :: CmmReg -> CmmExpr -> FCode ()

emitStore :: CmmExpr -> CmmExpr -> FCode ()

Assumes natural alignment.

emitStore' :: AlignmentSpec -> CmmExpr -> CmmExpr -> FCode ()

emitComment :: FastString -> FCode ()

emitTick :: CmmTickish -> FCode ()

emitUnwind :: [(GlobalReg, Maybe CmmExpr)] -> FCode ()

newTemp :: MonadUnique m => CmmType -> m LocalReg

getCmm :: FCode a -> FCode (a, CmmGroup)

aGraphToGraph :: CmmAGraphScoped -> FCode CmmGraph

getPlatform :: FCode Platform

getProfile :: FCode Profile

getCodeR :: FCode a -> FCode (a, CmmAGraph)

getCode :: FCode a -> FCode CmmAGraph

getCodeScoped :: FCode a -> FCode (a, CmmAGraphScoped)

Generate code into a fresh tick (sub-)scope and gather generated code

getHeapUsage :: (VirtualHpOffset -> FCode a) -> FCode a

getContext :: FCode SDocContext

mkCmmIfThenElse :: CmmExpr -> CmmAGraph -> CmmAGraph -> FCode CmmAGraph

mkCmmIfThen :: CmmExpr -> CmmAGraph -> FCode CmmAGraph

mkCmmIfGoto :: CmmExpr -> BlockId -> FCode CmmAGraph

mkCmmIfThenElse' :: CmmExpr -> CmmAGraph -> CmmAGraph -> Maybe Bool -> FCode CmmAGraph

mkCmmIfThen' :: CmmExpr -> CmmAGraph -> Maybe Bool -> FCode CmmAGraph

mkCmmIfGoto' :: CmmExpr -> BlockId -> Maybe Bool -> FCode CmmAGraph

mkCall :: CmmExpr -> (Convention, Convention) -> [CmmFormal] -> [CmmExpr] -> UpdFrameOffset -> [CmmExpr] -> FCode CmmAGraph

mkCmmCall :: CmmExpr -> [CmmFormal] -> [CmmExpr] -> UpdFrameOffset -> FCode CmmAGraph

forkClosureBody :: FCode () -> FCode ()

forkLneBody :: FCode a -> FCode a

forkAlts :: [FCode a] -> FCode [a]

forkAltPair :: FCode a -> FCode a -> FCode (a, a)

codeOnly :: FCode () -> FCode ()

type ConTagZ = Int

A *zero-indexed* constructor tag

data Sequel

A Sequel tells what to do with the result of this expression

Constructors

Return
Return result(s) to continuation found on the stack.

AssignTo

Fields

[LocalReg]

Put result(s) in these regs and fall through NB: no void arguments here

Bool

Should we adjust the heap pointer back to recover space that's unused on this path? We need to do this only if the expression may allocate (e.g. it's a foreign call or allocating primOp)

data ReturnKind

Constructors

AssignedDirectly

ReturnedTo BlockId ByteOff

withSequel :: Sequel -> FCode a -> FCode a

getSequel :: FCode Sequel

data SelfLoopInfo

Constructors

MkSelfLoopInfo

Fields

sli_id :: !Id

sli_arity :: !RepArity

always equal to idFunRepArity of sli_id, i.e. unarised arity, including void arguments

sli_registers :: ![LocalReg]

Excludes void arguments (LocalReg is never void)

sli_header_block :: !BlockId

setTickyCtrLabel :: CLabel -> FCode a -> FCode a

getTickyCtrLabel :: FCode CLabel

tickScope :: FCode a -> FCode a

Places blocks generated by the given code into a fresh (sub-)scope. This will make sure that Cmm annotations in our scope will apply to the Cmm blocks generated therein - but not the other way around.

getTickScope :: FCode CmmTickScope

The current tick scope. We will assign this to generated blocks.

withUpdFrameOff :: UpdFrameOffset -> FCode a -> FCode a

getUpdFrameOff :: FCode UpdFrameOffset

data HeapUsage

Constructors

HeapUsage

Fields

virtHp :: VirtualHpOffset

realHp :: VirtualHpOffset

type VirtualHpOffset = WordOff

initHpUsage :: HeapUsage

getHpUsage :: FCode HeapUsage

setHpUsage :: HeapUsage -> FCode ()

heapHWM :: HeapUsage -> VirtualHpOffset

setVirtHp :: VirtualHpOffset -> FCode ()

getVirtHp :: FCode VirtualHpOffset

setRealHp :: VirtualHpOffset -> FCode ()

getModuleName :: FCode Module

getState :: FCode CgState

setState :: CgState -> FCode ()

getSelfLoop :: FCode (Maybe SelfLoopInfo)

withSelfLoop :: SelfLoopInfo -> FCode a -> FCode a

getStgToCmmConfig :: FCode StgToCmmConfig

data CgIdInfo

Constructors

CgIdInfo

Fields

cg_id :: Id

Id that this is the info for

cg_lf :: LambdaFormInfo

cg_loc :: CgLoc

CmmExpr for the *tagged* value

getBinds :: FCode CgBindings

setBinds :: CgBindings -> FCode ()

data StgToCmmConfig

Constructors

StgToCmmConfig

Fields

stgToCmmProfile :: !Profile

Current profile

stgToCmmThisModule :: Module

The module being compiled. This field kept lazy for Cmm/Parser.y which preloads it with a panic

stgToCmmTmpDir :: !TempDir

Temp Dir for files used in compilation

stgToCmmContext :: !SDocContext

Context for StgToCmm phase

stgToCmmEmitDebugInfo :: !Bool

Whether we wish to output debug information

stgToCmmBinBlobThresh :: !(Maybe Word)

Threshold at which Binary literals (e.g. strings) are either dumped to a file and a CmmFileEmbed literal is emitted (over threshold), or become a CmmString Literal (under or at threshold). CmmFileEmbed is only supported with the NCG, thus a Just means two things: We have a threshold, and will be using the NCG. Conversely, a Nothing implies we are not using NCG and disables CmmFileEmbed. See Note [Embedding large binary blobs] in GHC.CmmToAsm.Ppr, and cgTopBinding in GHC.StgToCmm.

stgToCmmMaxInlAllocSize :: !Int

Max size, in bytes, of inline array allocations.

stgToCmmDoTicky :: !Bool

Ticky profiling enabled (cf -ticky)

stgToCmmTickyAllocd :: !Bool

True indicates ticky prof traces allocs of each named thing in addition to allocs _by_ that thing

stgToCmmTickyLNE :: !Bool

True indicates ticky uses name-specific counters for join-points (let-no-escape)

stgToCmmTickyDynThunk :: !Bool

True indicates ticky uses name-specific counters for dynamic thunks

stgToCmmTickyTag :: !Bool

True indicates ticky will count number of avoided tag checks by tag inference.

stgToCmmLoopification :: !Bool

Loopification enabled (cf -floopification)

stgToCmmAlignCheck :: !Bool

Insert alignment check (cf -falignment-sanitisation)

stgToCmmOptHpc :: !Bool

perform code generation for code coverage

stgToCmmFastPAPCalls :: !Bool

stgToCmmSCCProfiling :: !Bool

Check if cost-centre profiling is enabled

stgToCmmEagerBlackHole :: !Bool

stgToCmmInfoTableMap :: !Bool

true means generate C Stub for IPE map, See Note [Mapping Info Tables to Source Positions]

stgToCmmInfoTableMapWithFallback :: !Bool

Include info tables with fallback source locations in the info table map

stgToCmmInfoTableMapWithStack :: !Bool

Include info tables for STACK closures in the info table map

stgToCmmOmitYields :: !Bool

true means omit heap checks when no allocation is performed

stgToCmmOmitIfPragmas :: !Bool

true means don't generate interface programs (implied by -O0)

stgToCmmPIC :: !Bool

true if -fPIC

stgToCmmPIE :: !Bool

true if -fPIE

stgToCmmExtDynRefs :: !Bool

true if -fexternal-dynamic-refs, meaning generate code for linking against dynamic libraries

stgToCmmDoBoundsCheck :: !Bool

decides whether to check array bounds in StgToCmm.Prim or not

stgToCmmDoTagCheck :: !Bool

Verify tag inference predictions.

stgToCmmAllowBigArith :: !Bool

Allowed to emit larger than native size arithmetic (only LLVM and C backends)

stgToCmmAllowQuotRemInstr :: !Bool

Allowed to generate QuotRem instructions

stgToCmmAllowQuotRem2 :: !Bool

Allowed to generate QuotRem

stgToCmmAllowExtendedAddSubInstrs :: !Bool

Allowed to generate AddWordC, SubWordC, Add2, etc.

stgToCmmAllowIntMul2Instr :: !Bool

Allowed to generate IntMul2 instruction

stgToCmmTickyAP :: !Bool

Disable use of precomputed standard thunks.

stgToCmmVecInstrsErr :: Maybe String

Error (if any) to raise when vector instructions are used, see StgToCmm.Prim.checkVecCompatibility

stgToCmmAvx :: !Bool

check for Advanced Vector Extensions

stgToCmmAvx2 :: !Bool

check for Advanced Vector Extensions 2

stgToCmmAvx512f :: !Bool

check for Advanced Vector 512-bit Extensions

data CgState

Constructors

MkCgState

Fields

cgs_stmts :: CmmAGraph

cgs_tops :: OrdList CmmDecl

cgs_binds :: CgBindings

cgs_hp_usg :: HeapUsage

cgs_uniqs :: UniqSupply