GHC.StgToCmm.Sequel

Documentation

data Sequel

A Sequel tells what to do with the result of this expression

Constructors

Return
Return result(s) to continuation found on the stack.

AssignTo

Fields

[LocalReg]

Put result(s) in these regs and fall through NB: no void arguments here

Bool

Should we adjust the heap pointer back to recover space that's unused on this path? We need to do this only if the expression may allocate (e.g. it's a foreign call or allocating primOp)

data SelfLoopInfo

Constructors

MkSelfLoopInfo

Fields

sli_id :: !Id

sli_arity :: !RepArity

always equal to idFunRepArity of sli_id, i.e. unarised arity, including void arguments

sli_registers :: ![LocalReg]

Excludes void arguments (LocalReg is never void)

sli_header_block :: !BlockId