GHC.StgToJS.Apply

Description

Module that deals with expression application in JavaScript. In some cases we rely on pre-generated functions that are bundled with the RTS (see rtsApply).

Documentation

genApp :: HasDebugCallStack => ExprCtx -> Id -> [StgArg] -> G (JStat, ExprResult)

Generate an application of some args to an Id.

The case where args is null is common as it's used to generate the evaluation code for an Id.

rtsApply :: StgToJSConfig -> JStat

Pre-generated functions for fast Apply. These are bundled with the RTS.