GHC.StgToJS.CoreUtils

Description

Core utils

Documentation

isUnboxableCon :: DataCon -> Bool

can we unbox C x to x, only if x is represented as a Number

isUnboxable :: VarType -> Bool

one-constructor types with one primitive field represented as a JS Number can be unboxed

data SlotCount

Number of slots occupied by a PrimRep

Constructors

NoSlot

OneSlot

TwoSlots

slotCount :: SlotCount -> Int

Return SlotCount as an Int

varSize :: VarType -> Int

Number of slots occupied by a value with the given VarType

varSlotCount :: VarType -> SlotCount

typeSize :: Type -> Int

isVoid :: VarType -> Bool

isPtr :: VarType -> Bool

isSingleVar :: VarType -> Bool

isMultiVar :: VarType -> Bool

isMatchable :: [VarType] -> Bool

can we pattern match on these values in a case?

tyConVt :: HasDebugCallStack => TyCon -> [VarType]

idVt :: HasDebugCallStack => Id -> [VarType]

typeVt :: HasDebugCallStack => Type -> [VarType]

uTypeVt :: HasDebugCallStack => UnaryType -> VarType

primRepVt :: HasDebugCallStack => PrimRep -> VarType

typePrimRep' :: HasDebugCallStack => UnaryType -> [PrimRep]

tyConPrimRep' :: HasDebugCallStack => TyCon -> [PrimRep]

Find the primitive representation of a TyCon. Defined here to avoid module loops. Call this only on unlifted tycons.

kindPrimRep' :: HasDebugCallStack => SDoc -> Kind -> [PrimRep]

Take a kind (of shape TYPE rr) and produce the PrimReps of values of types of this kind.

primTypeVt :: HasDebugCallStack => Type -> VarType

argVt :: StgArg -> VarType

dataConType :: DataCon -> Type

isBoolDataCon :: DataCon -> Bool

fixedLayout :: [VarType] -> CILayout

stackSlotType :: Id -> VarType

idPrimReps :: Id -> [PrimRep]

typePrimReps :: Type -> [PrimRep]

primRepSize :: PrimRep -> SlotCount

assocPrimReps :: Outputable a => [PrimRep] -> [a] -> [(PrimRep, [a])]

Associate the given values to each RrimRep in the given order, taking into account the number of slots per PrimRep

assocIdPrimReps :: Outputable a => Id -> [a] -> [(PrimRep, [a])]

Associate the given values to the Id's PrimReps, taking into account the number of slots per PrimRep

assocIdExprs :: Id -> [JExpr] -> [TypedExpr]

Associate the given JExpr to the Id's PrimReps, taking into account the number of slots per PrimRep

might_be_a_function :: HasDebugCallStack => Type -> Bool

Return False only if we are *sure* it's a data type Look through newtypes etc as much as possible

mkArityTag :: Int -> Int -> Int

toTypeList :: [VarType] -> [Int]