GHC.StgToJS.DataCon

Description

Documentation

genCon :: ExprCtx -> DataCon -> [JExpr] -> G JStat

Generate a data constructor. Special handling for unboxed tuples

allocCon :: Ident -> DataCon -> CostCentreStack -> [JExpr] -> G JStat

Allocate a data constructor. Allocate in this context means bind the data constructor to to

allocUnboxedCon :: DataCon -> [JExpr] -> JExpr

Allocate an unboxed data constructor. If we have a bool we calculate the right value. If not then we expect a singleton list and unbox by converting ''C x' to x. NB. This function may panic.

allocDynamicE

Arguments

:: Bool
csInlineAlloc from StgToJSConfig

-> JExpr

-> [JExpr]

-> Maybe JExpr

-> JExpr

Allocate an entry function. See hs for the object layout.

allocDynamic :: StgToJSConfig -> Bool -> Ident -> JExpr -> [JExpr] -> Maybe JExpr -> JStat

Allocate a dynamic object