GHC.StgToJS.Expr

Description

Documentation

genExpr :: HasDebugCallStack => ExprCtx -> CgStgExpr -> G (JStat, ExprResult)

Evaluate an expression in the given expression context (continuation)

genEntryType :: HasDebugCallStack => [Id] -> G CIType

Generate the entry function types for identifiers. Note that this only returns either CIThunk or CIFun. Everything else (PAP Blackhole etc.) is filtered as not a RuntimeRepKinded type.

loadLiveFun :: [Id] -> G JStat

Given a set of Ids, bind each Id to the appropriate data fields in N registers. This assumes these data fields have already been populated in the registers. For the empty, singleton, and binary case use register 1, for any more use as many registers as necessary.

genStaticRefsRhs :: CgStgRhs -> G CIStatic

genStaticRefs :: LiveVars -> G CIStatic

genBody :: HasDebugCallStack => ExprCtx -> Id -> StgReg -> [Id] -> CgStgExpr -> G JStat

Generate the body of an object