GHC.StgToJS.ExprCtx

Description

Documentation

data ExprCtx

Context into which an expression is evaluated

initExprCtx :: Id -> ExprCtx

Initialize an expression context in the context of the given top-level binding Id

ctxAssertEvaluated :: Id -> ExprCtx -> ExprCtx

Add an Id to the known-evaluated set

ctxIsEvaluated :: ExprCtx -> Id -> Bool

Predicate: do we know for sure that the given Id is evaluated?

ctxSetSrcSpan :: RealSrcSpan -> ExprCtx -> ExprCtx

Set source location

ctxSrcSpan :: ExprCtx -> Maybe RealSrcSpan

Source location

ctxSetTop :: Id -> ExprCtx -> ExprCtx

Set top-level binding Id

ctxTarget :: ExprCtx -> [TypedExpr]

Target variables for the evaluated expression

ctxSetTarget :: [TypedExpr] -> ExprCtx -> ExprCtx

Set target

ctxEvaluatedIds :: ExprCtx -> UniqSet Id

Ids that we know to be evaluated (e.g. case binders when the expression to evaluate is in an alternative)

Let-no-escape

ctxClearLneFrame :: ExprCtx -> ExprCtx

Remove information about the current LNE frame

ctxUpdateLneFrame :: [(Id, Int)] -> [Id] -> ExprCtx -> ExprCtx

Update let-no-escape frame

ctxLneFrameVars :: ExprCtx -> [(Id, Int)]

Contents of current LNE frame

Variables and their index on the stack

ctxLneFrameSize :: ExprCtx -> Int

Cache the length of ctxLneFrameVars

ctxIsLneBinding :: ExprCtx -> Id -> Bool

Does the given Id correspond to a LNE binding

ctxIsLneLiveVar :: ExprCtx -> Id -> Bool

Does the given Id correspond to a LNE live var on the stack

ctxLneBindingStackSize :: ExprCtx -> Id -> Maybe Int

Return the LNE stack size associated to the given Id. Return Nothing when the Id doesn't correspond to a LNE binding.

ctxLneShrinkStack :: ExprCtx -> Int -> ExprCtx

Shrink the LNE stack to the given size