GHC.StgToJS.Ids

Description

Documentation

freshUnique :: G Int

Get fresh unique number

freshIdent :: G Ident

Get fresh local Ident of the form: h$$unit:module_uniq

makeIdentForId :: Id -> Maybe Int -> IdType -> Module -> Ident

Generate unique Ident for the given ID (uncached!)

The ident has the following forms:

global Id: h$unit:module.name[_num][_type_suffix] local Id: h$$unit:module.name[_num][_type_suffix]_uniq

Note that the string is z-encoded except for "_" delimiters.

Optional "_type_suffix" can be: - "_e" for IdEntry - "_con_e" for IdConEntry

Optional "_num" is passed as an argument to this function. It is used for Haskell Ids that require several JS variables: e.g. 64-bit numbers (Word64#, Int64#), Addr#, StablePtr#, unboxed tuples, etc.

cachedIdentForId :: Id -> Maybe Int -> IdType -> G Ident

Retrieve the cached Ident for the given Id if there is one. Otherwise make a new one with makeIdentForId and cache it.

Helpers for Idents

identForId :: Id -> G Ident

Retrieve default Ident for the given Id

identForIdN :: Id -> Int -> G Ident

Retrieve default Ident for the given Id with sub index

Some types, Word64, Addr#, unboxed tuple have more than one corresponding JS var, hence we use the sub index to identify each subpart / JS variable.

identsForId :: Id -> G [Ident]

Retrieve all the idents for the given Id.

identForEntryId :: Id -> G Ident

Retrieve entry Ident for the given Id

identForDataConEntryId :: Id -> G Ident

Retrieve datacon entry Ident for the given Id

Different name than the datacon wrapper.

identForDataConWorker :: DataCon -> G Ident

Retrieve datacon worker entry variable name for the given datacon

Helpers for variables

varForId :: Id -> G JExpr

Retrieve default variable name for the given Id

varForIdN :: Id -> Int -> G JExpr

Retrieve default variable name for the given Id with sub index

varsForId :: Id -> G [JExpr]

Retrieve all the JS vars for the given Id

varForEntryId :: Id -> G JExpr

Retrieve entry variable name for the given Id

varForDataConEntryId :: Id -> G JExpr

Retrieve datacon entry variable name for the given Id

varForDataConWorker :: DataCon -> G JExpr

Retrieve datacon worker entry variable name for the given datacon

declVarsForId :: Id -> G JStat

Declare all js vars for the id