GHC.StgToJS.Linker.Types

Description

Documentation

data JSLinkConfig

Constructors

JSLinkConfig

Fields

lcNoJSExecutables :: !Bool

Dont' build JS executables

lcNoHsMain :: !Bool

Don't generate Haskell main entry

lcNoRts :: !Bool

Don't dump the generated RTS

lcNoStats :: !Bool

Disable .stats file generation

lcForeignRefs :: !Bool

Dump .frefs (foreign references) files

lcCombineAll :: !Bool

Generate all.js (combined js) + wrappers

lcForceEmccRts :: !Bool

Force the link with the emcc rts. Use this if you plan to dynamically load wasm modules made from C files (e.g. in iserv).

lcLinkCsources :: !Bool

Link C sources (compiled to JS/Wasm) with Haskell code compiled to JS. This implies the use of the Emscripten RTS to load this code.

data LinkPlan

Constructors

LinkPlan

Fields

lkp_block_info :: Map Module LocatedBlockInfo

Block information

lkp_dep_blocks :: Set BlockRef

Blocks to link

lkp_archives :: !(Set FilePath)

Archives to load JS and Cc sources from (JS code corresponding to Haskell code is handled with blocks above)

lkp_objs_js :: !(Set FilePath)

JS objects to link

lkp_objs_cc :: !(Set FilePath)

Cc objects to link