GHC.StgToJS.Monad

Description

JS codegen state monad

Documentation

runG :: StgToJSConfig -> Module -> UniqFM Id CgStgExpr -> G a -> IO a

emitGlobal :: JStat -> G ()

emit a global (for the current module) toplevel statement

addDependency :: OtherSymb -> G ()

add a dependency on a particular symbol to the current group

emitToplevel :: JStat -> G ()

emit a top-level statement for the current binding group

emitStatic :: FastString -> StaticVal -> Maybe Ident -> G ()

emit static data for the binding group

emitClosureInfo :: ClosureInfo -> G ()

add closure info in our binding group. all heap objects must have closure info

emitForeign :: Maybe RealSrcSpan -> FastString -> Safety -> CCallConv -> [FastString] -> FastString -> G ()

assertRtsStat :: G JStat -> G JStat

getSettings :: G StgToJSConfig

globalOccs :: JStat -> G [GlobalOcc]

Return number of occurrences of every global id used in the given JStat. Sort by increasing occurrence count.

setGlobalIdCache :: GlobalIdCache -> G ()

getGlobalIdCache :: G GlobalIdCache

data GlobalOcc

Constructors

GlobalOcc

Fields

global_ident :: !Ident

global_id :: !Id

global_count :: !Word

Group

modifyGroup :: (GenGroupState -> GenGroupState) -> G ()

resetGroup :: G ()

start with a new binding group