GHC.StgToJS.Rts.Types

Description

Types and utility functions used in the JS RTS.

Documentation

traceRts :: StgToJSConfig -> JExpr -> JStat

Syntactic sugar, i.e., a Haskell function which generates useful JS code. Given a JExpr, ex, inject a trace statement on ex in the compiled JS program

assertRts :: ToJExpr a => StgToJSConfig -> JExpr -> a -> JStat

Syntactic sugar. Given a JExpr, ex which is assumed to be a predicate, and a message m, assert that 'not ex' is True, if not throw an exception in JS land with message m.

clName :: JExpr -> JExpr

name of the closure c

clTypeName :: JExpr -> JExpr

Type name of the closure c

stackFrameSize

Arguments

:: JExpr
assign frame size to this

-> JExpr
stack frame header function

-> JStat
size of the frame, including header

withRegs :: StgReg -> StgReg -> (StgReg -> JStat) -> JStat

Perform the computation f, on the range of registers bounded by start and end.