GHC.StgToJS.Symbols

Description

JS symbol generation

Documentation

intBS :: Int -> ByteString

Hexadecimal representation of an int

Used for the sub indices.

word64BS :: Word64 -> ByteString

Hexadecimal representation of a 64-bit word

Used for uniques. We could use base-62 as GHC usually does but this is likely faster.

unitModuleStringZ :: Module -> ByteString

Return z-encoded unit:module

moduleGlobalSymbol :: Module -> FastString

the global linkable unit of a module exports this symbol, depend on it to include that unit (used for cost centres)

moduleExportsSymbol :: Module -> FastString

mkJsSymbolBS :: Bool -> Module -> FastString -> ByteString

Make JS symbol corresponding to the given Haskell symbol in the given module

mkJsSymbol :: Bool -> Module -> FastString -> FastString

Make JS symbol corresponding to the given Haskell symbol in the given module

mkFreshJsSymbol :: Module -> Int -> FastString

Make JS symbol for given module and unique.

mkRawSymbol :: Bool -> FastString -> FastString

Make symbol "h$XYZ" or "h$$XYZ"

hddB :: ByteString

"h$$" constant string

hdB :: ByteString

"h$" constant string

hd :: JStgExpr

hdStr :: FastString

hdlB :: ByteString

hdApply :: JStgExpr

hdApplyStr :: FastString

hdMoveRegs2 :: FastString

hdPapGen :: JStgExpr

hdPapGenStr :: FastString

hdSetReg :: JStgExpr

hdSetRegStr :: FastString

hdGetReg :: JStgExpr

hdGetRegStr :: FastString

hdResetRegisters :: Ident

hdResetResultVars :: Ident

hdInitClosure :: FastString

hdRegs :: JStgExpr

hdRegsStr :: Ident

hdReturn :: JStgExpr

hdReturnStr :: Ident

hdStack :: JStgExpr

hdStackStr :: Ident

hdStackPtr :: JStgExpr

hdStackPtrStr :: Ident

hdBlackHoleTrap :: JStgExpr

hdBlackHoleTrapStr :: Ident

hdBlockOnBlackHoleStr :: FastString

hdBlackHoleLNE :: JStgExpr

hdBlackHoleLNEStr :: Ident

hdClosureTypeName :: JStgExpr

hdClosureTypeNameStr :: Ident

hdBh :: JStgExpr

hdBhStr :: FastString

hdBlackHole :: JStgExpr

hdBlackHoleStr :: Ident

hdUpdFrame :: JStgExpr

hdUpdFrameStr :: Ident

hdCSel :: JStgExpr

hdCSelStr :: FastString

hdEntry :: Ident

hdEntryStr :: FastString

hdApGen :: JStgExpr

hdApGenStr :: Ident

hdApGenFastStr :: Ident

hdLog :: JStgExpr

hdLogStr :: FastString

hdMkFunctionPtr :: JStgExpr

hdInitStatic :: JStgExpr

hdInitStaticStr :: Ident

hdHsSptInsert :: JStgExpr

hdCurrentThread :: JStgExpr

hdCurrentThreadStr :: Ident

hdWakeupThread :: FastString

hdPaps :: JStgExpr

hdPapsStr :: FastString

hdPapStr_ :: FastString

hdLazyEntryStr :: Ident

hdUnboxEntry :: JStgExpr

hdUnboxEntryStr :: Ident

hdMaskFrame :: JStgExpr

hdMaskFrameStr :: Ident

hdUnMaskFrameStr :: Ident

hdReturnF :: JStgExpr

hdReturnFStr :: Ident

hdResumeEntryStr :: Ident

hdFlushStdout :: JStgExpr

hdFlushStdoutStr :: Ident

hdFlushStdoutEntry :: JStgExpr

hdFlushStdoutEntryStr :: Ident

hdRunIOEntry :: JStgExpr

hdRunIOEntryStr :: Ident

hdReduce :: JStgExpr

hdReduceStr :: Ident

hdThrowStr :: FastString

hdRaiseAsyncFrame :: JStgExpr

hdRaiseAsyncFrameStr :: Ident

hdRaiseAsyncEntry :: JStgExpr

hdRaiseAsyncEntryStr :: Ident

hdRaiseEntry :: JStgExpr

hdRaiseEntryStr :: Ident

hdKeepAliveEntry :: JStgExpr

hdKeepAliveEntryStr :: Ident

hdSelect2Return :: JStgExpr

hdSelect2ReturnStr :: Ident

hdSelect2Entry :: JStgExpr

hdSelect2EntryStr :: Ident

hdSelect1Ret :: JStgExpr

hdSelect1RetStr :: Ident

hdSelect1EntryStr :: Ident

hdStaticThunkStr :: FastString

hdStaticThunksStr :: Ident

hdStaticThunksArrStr :: Ident

hdCAFsStr :: Ident

hdCAFsResetStr :: Ident

hdUpdThunkEntryStr :: Ident

hdAp3EntryStr :: Ident

hdAp2EntryStr :: Ident

hdAp1EntryStr :: Ident

hdDataToTagEntryStr :: Ident

hdTagToEnum :: FastString

hdCatchEntryStr :: Ident

hdNoopStr :: Ident

hdNoopEntry :: JStgExpr

hdNoopEntryStr :: Ident

hdC0 :: JStgExpr

hdC :: JStgExpr

hdC0Str :: Ident

hdCStr :: Ident

hdData2Entry :: Ident

hdData1Entry :: Ident

hdTrueEntry :: Ident

hdFalseEntry :: Ident

hdDoneMainEntry :: JStgExpr

hdDoneMainEntryStr :: Ident

hdDoneMain :: JStgExpr

hdDone :: Ident

hdExitProcess :: FastString

hdTraceAlloc :: FastString

hdDebugAllocNotifyAlloc :: FastString

hdRtsTraceForeign :: Ident

hdRtsProfiling :: Ident

hdCtFun :: Ident

hdCtCon :: Ident

hdCtThunk :: Ident

hdCtPap :: Ident

hdCtBlackhole :: Ident

hdCtStackFrame :: Ident

hdCtVtPtr :: Ident

hdVtVoid :: Ident

hdVtInt :: Ident

hdVtDouble :: Ident

hdVtLong :: Ident

hdVtAddr :: Ident

hdVtObj :: Ident

hdVtArr :: Ident

hdLoads :: Array Int Ident

hdAp00 :: JStgExpr

hdAp00Str :: Ident

hdAp00FastStr :: FastString

hdAp11Fast :: FastString

hdAp10 :: JStgExpr

hdAp33FastStr :: FastString

hdAp22FastStr :: FastString

hdNewByteArrayStr :: FastString

hdCopyMutableByteArrayStr :: FastString

hdCheckOverlapByteArrayStr :: FastString

hdShrinkMutableCharArrayStr :: FastString

hdTraceEventStr :: FastString

hdTraceEventBinaryStr :: FastString

hdTraceMarkerStr :: FastString

hdThrowJSException :: JStgExpr

hdUnboxFFIResult :: JStgExpr

hdUnboxFFIResultStr :: Ident

hdMkForeignCallback :: JStgExpr

hdTraceForeign :: JStgExpr

hdBuildObject :: JStgExpr

hdBuildObjectStr :: FastString

hdCallDynamicStr :: FastString

except :: JStgExpr

exceptStr :: Ident

excepStr :: FastString

mv :: FastString

lngth :: FastString

len :: FastString

only for byte arrays. This is a JS byte array method

slice :: FastString

this :: JStgExpr

arr :: FastString

dv :: FastString

d1 :: JStgExpr

d2 :: JStgExpr

d3 :: JStgExpr

d1Str :: FastString

d2Str :: FastString

d3Str :: FastString

getInt16 :: FastString

getUint16 :: FastString

getInt32 :: FastString

getUint32 :: FastString

getFloat32 :: FastString

getFloat64 :: FastString

setInt16 :: FastString

setUint16 :: FastString

setInt32 :: FastString

setUint32 :: FastString

setFloat32 :: FastString

setFloat64 :: FastString

i3 :: FastString

u8 :: FastString

u1 :: FastString

f6 :: FastString

f3 :: FastString

val :: FastString

label :: FastString

mask :: FastString

unMask :: FastString

resume :: FastString

f :: FastString

n :: FastString

hasOwnProperty :: FastString

hdCollectProps :: FastString

replace :: FastString

substring :: FastString

trace :: FastString

apply :: FastString

hdMVar :: JStgExpr

hdMVarStr :: FastString

hdTakeMVar :: JStgExpr

hdTakeMVarStr :: FastString

hdTryTakeMVarStr :: FastString

hdPutMVarStr :: FastString

hdTryPutMVarStr :: FastString

hdNewTVar :: FastString

hdReadTVar :: FastString

hdReadTVarIO :: FastString

hdWriteTVar :: FastString

hdReadMVarStr :: FastString

hdStmRemoveBlockedThreadStr :: FastString

hdStmStartTransactionStr :: FastString

hdAtomicallyEntry :: JStgExpr

hdAtomicallyEntryStr :: Ident

hdAtomicallyStr :: FastString

hdStgResumeRetryEntry :: JStgExpr

hdStgResumeRetryEntryStr :: Ident

hdStmCommitTransactionStr :: FastString

hdStmValidateTransactionStr :: FastString

hdStmCatchRetryEntry :: JStgExpr

hdStmCatchRetryEntryStr :: Ident

hdStmRetryStr :: FastString

hdStmCatchRetryStr :: FastString

hdStmCatchEntry :: JStgExpr

hdCatchStmStr :: FastString

hdStmCatchEntryStr :: Ident

hdRetryInterrupted :: JStgExpr

hdRetryInterruptedStr :: Ident

hdMaskUnintFrame :: JStgExpr

hdMaskUnintFrameStr :: Ident

hdReschedule :: JStgExpr

hdRescheduleStr :: Ident

hdRestoreThread :: JStgExpr

hdRestoreThreadStr :: Ident

hdFinishedThread :: FastString

hdPrimOpStr :: FastString

wrapperColonStr :: FastString

hdInternalExceptionTypeDivZero :: JStgExpr

hdInternalExceptionTypeOverflow :: JStgExpr

hdInternalExceptionTypeUnderflow :: JStgExpr

hdInternalExceptionControlExceptionBaseNonTermination :: JStgExpr

hdGhcInternalIOHandleFlush :: JStgExpr

hdGhcInternalIOHandleFDStdout :: JStgExpr

hdGhcInternalJSPrimValConEntryStr :: FastString

hdBuildCCSPtrStr :: FastString

hdClearCCSStr :: FastString

hdRestoreCCSStr :: FastString

hdSetCcsEntry :: JStgExpr

hdSetCcsEntryStr :: Ident

ccStr :: FastString

unknown :: FastString

typeof :: FastString

hdRawStr :: FastString

throwStr :: FastString

hdCheckObj :: JStgExpr

console :: JStgExpr

consoleStr :: FastString

arguments :: JStgExpr

argumentsStr :: FastString

hdReportHeapOverflow :: JStgExpr

hdReportHeapOverflowStr :: Ident

hdReportStackOverflow :: JStgExpr

hdReportStackOverflowStr :: Ident

hdDumpRes :: JStgExpr

hdDumpResStr :: Ident

ghcjsArray :: FastString

hdCompactSize :: FastString

hdCompactAddWithSharing :: FastString

hdCompactAdd :: FastString

hdCompactFixupPointers :: FastString

hdCompactAllocateBlock :: FastString

hdCompactGetNextBlock :: FastString

hdCompactGetFirstBlock :: FastString

hdCompactContainsAny :: FastString

hdCompactContains :: FastString

hdCompactResize :: FastString

hdCompactNew :: FastString

hdStableNameInt :: FastString

hdMakeStableName :: FastString

hdDeRefStablePtr :: FastString

hdStablePtrBuf :: JStgExpr

hdMakeStablePtrStr :: FastString

hdKeepAlive :: FastString

hdFinalizeWeak :: FastString

hdMakeWeakNoFinalizer :: FastString

hdMakeWeak :: FastString

hdGetThreadLabel :: FastString

hdListThreads :: FastString

hdThreadStatus :: FastString

hdYield :: FastString

hdKillThread :: FastString

hdFork :: FastString

hdWaitWrite :: FastString

hdWaitRead :: FastString

hdDelayThread :: FastString

hdCatchStr :: FastString

hdMaskAsyncStr :: FastString

hdMaskUnintAsyncStr :: FastString

hdUnmaskAsyncStr :: FastString

hdMutVarStr :: FastString

hdAtomicModifyMutVar2Str :: FastString

hdAtomicModifyMutVarStr :: FastString

hdComparePointerStr :: FastString

hdCompareByteArraysStr :: FastString

hdResizeMutableByteArrayStr :: FastString

hdShrinkMutableByteArrayStr :: FastString

hdCopyMutableArrayStr :: FastString

hdNewArrayStr :: FastString

hdSliceArrayStr :: FastString

hdDecodeFloatIntStr :: FastString

hdCastFloatToWord32Str :: FastString

hdCastWord32ToFloatStr :: FastString

hdDecodeDouble2IntStr :: FastString

hdDecodeDoubleInt64Str :: FastString

hdCastDoubleToWord64Str :: FastString

hdCastWord64ToDoubleStr :: FastString

hdReverseWordStr :: FastString

hdClz8Str :: FastString

hdClz16Str :: FastString

hdClz32Str :: FastString

hdClz64Str :: FastString

hdCtz8Str :: FastString

hdCtz16Str :: FastString

hdCtz32Str :: FastString

hdCtz64Str :: FastString

hdBSwap64Str :: FastString

hdPExit8Str :: FastString

hdPExit16Str :: FastString

hdPExit32Str :: FastString

hdPExit64Str :: FastString

hdPDep8Str :: FastString

hdPDep16Str :: FastString

hdPDep32Str :: FastString

hdPDep64Str :: FastString

hdPopCntTab :: JStgExpr

hdPopCnt32Str :: FastString

hdPopCnt64Str :: FastString

hdQuotRem2Word32Str :: FastString

hdQuotRemWord32Str :: FastString

hdRemWord32Str :: FastString

hdQuotWord32Str :: FastString

hdMul2Word32Str :: FastString

hdMulImulStr :: FastString

hdWordAdd2 :: FastString

hdHsPlusWord64Str :: FastString

hdHsMinusWord64Str :: FastString

hdHsTimesWord64Str :: FastString

hdHsQuotWord64Str :: FastString

hdHsRemWord64Str :: FastString

hdHsUncheckedShiftRWord64Str :: FastString

hdHsUncheckedShiftLWord64Str :: FastString

hdHsPlusInt64Str :: FastString

hdHsMinusInt64Str :: FastString

hdHsTimesInt64Str :: FastString

hdHsQuotInt64Str :: FastString

hdHsRemInt64Str :: FastString

hdHsUncheckedShiftLLInt64Str :: FastString

hdHsUncheckedShiftRAInt64Str :: FastString

hdHsUncheckedShiftRLInt64Str :: FastString

hdHsTimesInt2Str :: FastString

hdEncodeModifiedUtf8Str :: FastString

hdRawStringDataStr :: FastString

hdPStr :: FastString

hdDStr :: FastString

hdDiStr :: FastString

hdStcStr :: FastString

hdStlStr :: FastString

hdStiStr :: FastString

hdStrStr :: FastString