GHC.StgToJS.Types

Description

Module that holds the Types required for the StgToJS pass

Documentation

type G = StateT GenState IO

A State monad over IO holding the generator state.

data GenState

The JS code generator state

Constructors

GenState

Fields

gsSettings :: !StgToJSConfig

codegen settings, read-only

gsModule :: !Module

current module

gsId :: !FastMutInt

unique number for the id generator

gsIdents :: !IdCache

hash consing for identifiers from a Unique

gsUnfloated :: !(UniqFM Id CgStgExpr)

unfloated arguments

gsGroup :: GenGroupState

state for the current binding group

gsGlobal :: [JStgStat]

global (per module) statements (gets included when anything else from the module is used)

data GenGroupState

The JS code generator state relevant for the current binding group

Constructors

GenGroupState

Fields

ggsToplevelStats :: [JStgStat]

extra toplevel statements for the binding group

ggsClosureInfo :: [ClosureInfo]

closure metadata (info tables) for the binding group

ggsStatic :: [StaticInfo]

static (CAF) data in our binding group

ggsStack :: [StackSlot]

stack info for the current expression

ggsStackDepth :: Int

current stack depth

ggsExtraDeps :: Set OtherSymb

extra dependencies for the linkable unit that contains this group

ggsGlobalIdCache :: GlobalIdCache

ggsForeignRefs :: [ForeignJSRef]

data StgToJSConfig

The Configuration record for the StgToJS pass

Constructors

StgToJSConfig

Fields

csInlinePush :: !Bool

csInlineBlackhole :: !Bool

csInlineLoadRegs :: !Bool

csInlineEnter :: !Bool

csInlineAlloc :: !Bool

csPrettyRender :: !Bool

csTraceRts :: !Bool

csAssertRts :: !Bool

csBoundsCheck :: !Bool

csDebugAlloc :: !Bool

csTraceForeign :: !Bool

csProf :: !Bool

Profiling enabled

csRuntimeAssert :: !Bool

Enable runtime assertions settings

csContext :: !SDocContext

csLinkerConfig :: !LinkerConfig

Emscripten linker

data ClosureInfo

Closure info table

Constructors

ClosureInfo

Fields

ciVar :: Ident

entry code identifier: infotable fields are stored as properties of this function

ciRegs :: CIRegs

size of the payload (in number of JS values)

ciName :: FastString

friendly name for printing

ciLayout :: CILayout

heap/stack layout of the object

ciType :: CIType

type of the object, with extra info where required

ciStatic :: CIStatic

static references of this object

data CIRegs

Closure information, ClosureInfo, registers

Constructors

CIRegsUnknown
A value witnessing a state of unknown registers

CIRegs

Fields

ciRegsSkip :: Int

unused registers before actual args start

ciRegsTypes :: [JSRep]

args

data CILayout

Closure Information, ClosureInfo, layout

Constructors

CILayoutVariable
layout stored in object itself, first position from the start

CILayoutUnknown
fixed size, but content unknown (for example stack apply frame)

Fields

layoutSize :: !Int

closure size in array positions, including entry

CILayoutFixed
whole layout known

Fields

layoutSize :: !Int

closure size in array positions, including entry

layout :: [JSRep]

The list of JSReps to layout

data CIType

The type of ClosureInfo

Constructors

CIFun

Fields

citArity :: !Int

function arity

citRegs :: !Int

number of registers for the args

CIThunk
The closure is a THUNK

CICon
The closure is a Constructor

Fields

citConstructor :: !Int

CIPap
The closure is a Partial Application

CIBlackhole
The closure is a black hole

CIStackFrame
The closure is a stack frame

newtype CIStatic

Static references that must be kept alive

Constructors

CIStaticRefs

Fields

staticRefs :: [FastString]

data JSRep

JS primitive representations

Constructors

PtrV
pointer = reference to heap object (closure object), lifted or not. Can also be some RTS object (e.g. TVar#, MVar#, MutVar#, Weak#)

VoidV
no fields

DoubleV
A Double: one field

IntV
An Int (32bit because JS): one field

LongV
A Long: two fields one for the upper 32bits, one for the lower (NB: JS is little endian)

AddrV
a pointer not to the heap: two fields, array + index

ObjV
some JS object, user supplied, be careful around these, can be anything

ArrV
boxed array

data IdType

The type of identifiers. These determine the suffix of generated functions in JS Land. For example, the entry function for the Just constructor is a IdConEntry which compiles to: function h$ghczminternalZCGHCziInternalziMaybeziJust_con_e() { return h$rs() }; which just returns whatever the stack point is pointing to. Whereas the entry function to Just is an IdEntry and does the work. It compiles to: function h$ghczminternalZCGHCziInternalziMaybeziJust_e() { var h$$baseZCGHCziMaybezieta_8KXnScrCjF5 = h$r2; h$r1 = h$c1(h$ghczminternalZCGHCziInternalziMaybeziJust_con_e, h$$ghczminternalZCGHCziInternalziMaybezieta_8KXnScrCjF5); return h$rs(); }; Which loads some payload from register 2, and applies the Constructor Entry function for the Just to the payload, returns the result in register 1 and returns whatever is on top of the stack

Constructors

IdPlain
A plain identifier for values, no suffix added

IdEntry
An entry function, suffix = "_e" in makeIdentForId

IdConEntry
A Constructor entry function, suffix = "_con_e" in makeIdentForId

data IdKey

Keys to differentiate Ident's in the ID Cache

Constructors

IdKey !Word64 !Int !IdType

data OtherSymb

Some other symbol

Constructors

OtherSymb !Module !FastString

newtype IdCache

The identifier cache indexed on IdKey local to a module

Constructors

IdCache (Map IdKey Ident)

newtype GlobalIdCache

The global Identifier Cache

Constructors

GlobalIdCache (UniqFM Ident (IdKey, Id))

data StackSlot

A Stack Slot is either known or unknown. We avoid maybe here for more strictness.

Constructors

SlotId !Id !Int

SlotUnknown

data StaticInfo

Constructors

StaticInfo

Fields

siVar :: !FastString

global object

siVal :: !StaticVal

static initialization

siCC :: !(Maybe Ident)

optional CCS name

data StaticVal

Constructors

StaticFun !FastString [StaticArg]
heap object for function

StaticThunk !(Maybe (FastString, [StaticArg]))
heap object for CAF (field is Nothing when thunk is initialized in an alternative way, like string thunks through h$str)

StaticUnboxed !StaticUnboxed
unboxed constructor (Bool, Int, Double etc)

StaticData !FastString [StaticArg]
regular datacon app

StaticList [StaticArg] (Maybe FastString)
list initializer (with optional tail)

data StaticUnboxed

Constructors

StaticUnboxedBool !Bool

StaticUnboxedInt !Integer

StaticUnboxedDouble !SaneDouble

StaticUnboxedString !ByteString

StaticUnboxedStringOffset !ByteString

data StaticArg

Static Arguments. Static Arguments are things that are statically allocated, i.e., they exist at program startup. These are static heap objects or literals or things that have been floated to the top level binding by ghc.

Constructors

StaticObjArg !FastString
reference to a heap object

StaticLitArg !StaticLit
literal

StaticConArg !FastString [StaticArg]
unfloated constructor

data StaticLit

A Static literal value

Constructors

BoolLit !Bool

IntLit !Integer

NullLit

DoubleLit !SaneDouble

StringLit !FastString

BinLit !ByteString

LabelLit !Bool !FastString
is function pointer, label (also used for string / binary init)

data ForeignJSRef

A foreign reference to some JS code

Constructors

ForeignJSRef

Fields

foreignRefSrcSpan :: !FastString

foreignRefPattern :: !FastString

foreignRefSafety :: !Safety

foreignRefCConv :: !CCallConv

foreignRefArgs :: ![FastString]

foreignRefResult :: !FastString

data LinkableUnit

data used to generate one ObjBlock in our object file

Constructors

LinkableUnit

Fields

luObjBlock :: ObjBlock

serializable unit info

luIdExports :: [Id]

exported names from haskell identifiers

luOtherExports :: [FastString]

other exports

luIdDeps :: [Id]

identifiers this unit depends on

luPseudoIdDeps :: [Unique]

pseudo-id identifiers this unit depends on (fixme)

luOtherDeps :: [OtherSymb]

symbols not from a haskell id that this unit depends on

luRequired :: Bool

always link this unit

luForeignRefs :: [ForeignJSRef]

data ObjBlock

one toplevel block in the object file

Constructors

ObjBlock

Fields

oiSymbols :: ![FastString]

toplevel symbols (stored in index)

oiClInfo :: ![ClosureInfo]

closure information of all closures in block

oiStatic :: ![StaticInfo]

static closure data

oiStat :: JStat

the code

oiRaw :: !ByteString

raw JS code

oiFExports :: ![ExpFun]

oiFImports :: ![ForeignJSRef]

data ExpFun

Constructors

ExpFun

Fields

isIO :: !Bool

args :: [JSFFIType]

result :: !JSFFIType

data JSFFIType

Types of FFI values

Constructors

Int8Type

Int16Type

Int32Type

Int64Type

Word8Type

Word16Type

Word32Type

Word64Type

DoubleType

ByteArrayType

PtrType

RefType

data TypedExpr

Typed expression

Constructors

TypedExpr

Fields

typex_typ :: !PrimRep

typex_expr :: [JStgExpr]

data PrimRes

A Primop result is either an inlining of some JS payload, or a primitive call to a JS function defined in Shim files in base.

Constructors

PrimInline JStgStat
primop is inline, result is assigned directly

PRPrimCall JStgStat
primop is async call, primop returns the next function to run. result returned to stack top in registers

data ExprResult

Constructors

ExprCont

ExprInline

newtype ExprValData

Constructors

ExprValData [JStgExpr]

data ClosureType

A Closure is one of six types

Constructors

Thunk
The closure is a THUNK

Fun
The closure is a Function

Pap
The closure is a Partial Application

Con
The closure is a Constructor

Blackhole
The closure is a Blackhole

StackFrame
The closure is a stack frame

ctNum :: ClosureType -> Int

Convert ClosureType to an Int

closureB :: ByteString

closureNames :: Array ClosureType Ident

ctJsName :: ClosureType -> String

Convert ClosureType to a String

data ThreadStatus

A thread is in one of 4 states

Constructors

Running
The thread is running

Blocked
The thread is blocked

Finished
The thread is done

Died
The thread has died

threadStatusNum :: ThreadStatus -> Int

Convert the status of a thread in JS land to an Int

threadStatusJsName :: ThreadStatus -> String

convert the status of a thread in JS land to a string