GHC.StgToJS.Utils

Documentation

assignCoerce1 :: [TypedExpr] -> [TypedExpr] -> JStgStat

Assign first expr only (if it exists), performing coercions between some PrimReps (e.g. StablePtr# and Addr#).

assignToExprCtx :: ExprCtx -> [JStgExpr] -> JStgStat

fixedLayout :: [JSRep] -> CILayout

assocIdExprs :: Id -> [JStgExpr] -> [TypedExpr]

Associate the given JExpr to the Id's PrimReps, taking into account the number of slots per PrimRep

Unboxable datacon

isUnboxableCon :: DataCon -> Bool

can we unbox C x to x, only if x is represented as a Number

isUnboxable :: JSRep -> Bool

one-constructor types with one primitive field represented as a JS Number can be unboxed

isBoolDataCon :: DataCon -> Bool

JSRep

slotCount :: SlotCount -> Int

Return SlotCount as an Int

varSize :: JSRep -> Int

Number of slots occupied by a value with the given JSRep

typeSize :: Type -> Int

isVoid :: JSRep -> Bool

isMultiVar :: JSRep -> Bool

idJSRep :: HasDebugCallStack => Id -> [JSRep]

typeJSRep :: HasDebugCallStack => Type -> [JSRep]

unaryTypeJSRep :: HasDebugCallStack => UnaryType -> JSRep

primRepToJSRep :: HasDebugCallStack => PrimRep -> JSRep

primOrVoidRepToJSRep :: HasDebugCallStack => PrimOrVoidRep -> JSRep

stackSlotType :: Id -> JSRep

primRepSize :: PrimRep -> SlotCount

mkArityTag :: Int -> Int -> Int

References and Ids

exprRefs :: UniqFM Id CgStgExpr -> CgStgExpr -> Set Id

hasExport :: CgStgBinding -> Bool

collectTopIds :: CgStgBinding -> [Id]

collectIds :: UniqFM Id CgStgExpr -> CgStgBinding -> [Id]

Live variables

type LiveVars = DVarSet

liveStatic :: LiveVars -> LiveVars

liveVars :: LiveVars -> LiveVars

stgRhsLive :: CgStgRhs -> LiveVars

stgExprLive :: Bool -> CgStgExpr -> LiveVars

isUpdatableRhs :: CgStgRhs -> Bool

stgLneLive' :: CgStgBinding -> [Id]

stgLneLiveExpr :: CgStgRhs -> [Id]

isInlineExpr :: CgStgExpr -> Bool

returns True if the expression is definitely inline