GHC.SysTools.Cpp

Documentation

doCpp :: Logger -> TmpFs -> DynFlags -> UnitEnv -> CppOpts -> FilePath -> FilePath -> IO ()

Run either the Haskell preprocessor, JavaScript preprocessor or the C preprocessor, as per the CppOpts passed. See Note [Preprocessing invocations].

UnitEnv is needed to compute MIN_VERSION macros

If you change the macros defined by this function make sure to update the user guide.

data CppOpts

Constructors

CppOpts

Fields

sourceCodePreprocessor :: !SourceCodePreprocessor

cppLinePragmas :: !Bool

Enable generation of LINE pragmas

getGhcVersionPathName :: DynFlags -> UnitEnv -> IO FilePath

Find out path to ghcversion.h file

applyCDefs :: DefunctionalizedCDefs -> Logger -> DynFlags -> IO [String]

offsetIncludePaths :: DynFlags -> IncludeSpecs -> IncludeSpecs