GHC.SysTools.Elf

Documentation

readElfSectionByName :: Logger -> DynFlags -> ByteString -> String -> IO (Maybe ByteString)

Given a section name, read its contents as a ByteString.

If the section isn't found or if there is any parsing error, we return Nothing

readElfNoteAsString :: Logger -> DynFlags -> FilePath -> String -> String -> IO (Maybe String)

read a Note as a String

If you try to read a note from a section which does not support the Note format, the parsing is likely to fail and Nothing will be returned

makeElfNote :: Platform -> String -> String -> Word32 -> String -> SDoc

Generate the GAS code to create a Note section

Header fields for notes are 32-bit long (see Note [ELF specification]).