GHC.SysTools.Process

Documentation

readCreateProcessWithExitCode'

Arguments

:: CreateProcess

-> IO (ExitCode, String)
stdout

getGccEnv :: [Option] -> IO (Maybe [(String, String)])

runSomething :: Logger -> String -> String -> [Option] -> IO ()

runSomethingResponseFile :: Logger -> TmpFs -> TempDir -> ([String] -> [String]) -> String -> String -> [Option] -> Maybe [(String, String)] -> IO ()

Run a command, placing the arguments in an external response file.

This command is used in order to avoid overlong command line arguments on Windows. The command line arguments are first written to an external, temporary response file, and then passed to the linker via @filepath. response files for passing them in. See:

https://gcc.gnu.org/wiki/Response_Files https://gitlab.haskell.org/ghc/ghc/issues/10777

runSomethingFiltered :: Logger -> ([String] -> [String]) -> String -> String -> [Option] -> Maybe FilePath -> Maybe [(String, String)] -> IO ()

runSomethingWith :: Logger -> String -> String -> [Option] -> ([String] -> IO (ExitCode, a)) -> IO a