GHC.SysTools.Tasks

Documentation

runUnlit :: Logger -> DynFlags -> [Option] -> IO ()

augmentImports :: DynFlags -> [FilePath] -> [FilePath]

Prepend the working directory to the search path. Note [Filepaths and Multiple Home Units]

cc_filter :: String -> String

Discard some harmless warnings from gcc that we can't turn off

data SourceCodePreprocessor

See the Note [Preprocessing invocations]

Constructors

SCPCpp
Use the ordinary C preprocessor

SCPHsCpp
Use the Haskell C preprocessor (don't remove C comments, don't break on names including single quotes)

SCPJsCpp
Use the JavaScript preprocessor (don't remove jsdoc and multiline comments)

SCPCmmCpp
Use the C-- preprocessor (don't emit debug information)

runSourceCodePreprocessor :: Logger -> TmpFs -> DynFlags -> SourceCodePreprocessor -> [Option] -> IO ()

Run source code preprocessor. See also Note [Preprocessing invocations] in GHC.SysTools.Cpp

runPp :: Logger -> DynFlags -> [Option] -> IO ()

runCc :: Maybe ForeignSrcLang -> Logger -> TmpFs -> DynFlags -> [Option] -> IO ()

Run compiler of C-like languages and raw objects (such as gcc or clang).

isContainedIn :: String -> String -> Bool

askLd :: Logger -> DynFlags -> [Option] -> IO String

Run the linker with some arguments and return the output

runAs :: Logger -> DynFlags -> [Option] -> IO ()

runLlvmOpt :: Logger -> DynFlags -> [Option] -> IO ()

Run the LLVM Optimiser

runLlvmLlc :: Logger -> DynFlags -> [Option] -> IO ()

Run the LLVM Compiler

runLlvmAs :: Logger -> DynFlags -> [Option] -> IO ()

Run the LLVM Assembler

runEmscripten :: Logger -> DynFlags -> [Option] -> IO ()

figureLlvmVersion :: Logger -> DynFlags -> IO (Maybe LlvmVersion)

Figure out which version of LLVM we are running this session

runMergeObjects :: Logger -> TmpFs -> DynFlags -> [Option] -> IO ()

runAr :: Logger -> DynFlags -> Maybe FilePath -> [Option] -> IO ()

askOtool :: Logger -> ToolSettings -> Maybe FilePath -> [Option] -> IO String

runInstallNameTool :: Logger -> ToolSettings -> [Option] -> IO ()

runRanlib :: Logger -> DynFlags -> [Option] -> IO ()

runWindres :: Logger -> DynFlags -> [Option] -> IO ()