GHC.Tc.Deriv.Functor

Description

The deriving code for the Functor, Foldable, and Traversable classes

Documentation

data FFoldType a

Constructors

FT

Fields

ft_triv :: a

Does not contain variable

ft_var :: a

The variable itself

ft_co_var :: a

The variable itself, contravariantly

ft_fun :: a -> a -> a

Function type

ft_tup :: TyCon -> [a] -> a

Tuple type. The [a] is the result of folding over the arguments of the tuple.

ft_ty_app :: Type -> Type -> a -> a

Type app, variable only in last argument. The two Types are the function and argument parts of fun_ty arg_ty, respectively.

ft_bad_app :: a

Type app, variable other than in last argument

ft_forall :: TcTyVar -> a -> a

Forall type

functorLikeTraverse

Arguments

:: forall a. TyVar
Variable to look for

-> FFoldType a
How to fold

-> Type
Type to process

-> a

deepSubtypesContaining :: TyVar -> Type -> [TcType]

foldDataConArgs :: FFoldType a -> DataCon -> [a]

gen_Functor_binds :: SrcSpan -> TyCon -> [Type] -> (LHsBinds GhcPs, BagDerivStuff)

gen_Foldable_binds :: SrcSpan -> TyCon -> [Type] -> (LHsBinds GhcPs, BagDerivStuff)

gen_Traversable_binds :: SrcSpan -> TyCon -> [Type] -> (LHsBinds GhcPs, BagDerivStuff)