GHC.Tc.Errors.Hole.FitTypes

Documentation

data TypedHole

Constructors

TypedHole

Fields

th_relevant_cts :: Cts

Any relevant Cts to the hole

th_implics :: [Implication]

The nested implications of the hole with the innermost implication first.

th_hole :: Maybe Hole

The hole itself, if available. Only for debugging.

data HoleFit

HoleFit is the type we use for valid hole fits. It contains the element that was checked, the Id of that element as found by tcLookup, and the refinement level of the fit, which is the number of extra argument holes that this fit uses (e.g. if hfRefLvl is 2, the fit is for `Id _ _`).

Constructors

HoleFit

Fields

hfId :: Id

The elements id in the TcM

hfCand :: HoleFitCandidate

The candidate that was checked.

hfType :: TcType

The type of the id, possibly zonked.

hfRefLvl :: Int

The number of holes in this fit.

hfWrap :: [TcType]

The wrapper for the match.

hfMatches :: [TcType]

What the refinement variables got matched with, if anything

hfDoc :: Maybe HsDocString

Documentation of this HoleFit, if available.

RawHoleFit SDoc
A fit that is just displayed as is. Here so thatHoleFitPlugins can inject any fit they want.

data HoleFitCandidate

HoleFitCandidates are passed to hole fit plugins and then checked whether they fit a given typed-hole.

Constructors

IdHFCand Id

NameHFCand Name

GreHFCand GlobalRdrElt

type CandPlugin = TypedHole -> [HoleFitCandidate] -> TcM [HoleFitCandidate]

A plugin for modifying the candidate hole fits *before* they're checked.

type FitPlugin = TypedHole -> [HoleFit] -> TcM [HoleFit]

A plugin for modifying hole fits *after* they've been found.

data HoleFitPlugin

A HoleFitPlugin is a pair of candidate and fit plugins.

Constructors

HoleFitPlugin

Fields

candPlugin :: CandPlugin

fitPlugin :: FitPlugin

data HoleFitPluginR

HoleFitPluginR adds a TcRef to hole fit plugins so that plugins can track internal state. Note the existential quantification, ensuring that the state cannot be modified from outside the plugin.

Constructors

forall s. HoleFitPluginR

Fields

hfPluginInit :: TcM (TcRef s)

Initializes the TcRef to be passed to the plugin

hfPluginRun :: TcRef s -> HoleFitPlugin

The function defining the plugin itself

hfPluginStop :: TcRef s -> TcM ()

Cleanup of state, guaranteed to be called even on error

hfIsLcl :: HoleFit -> Bool

pprHoleFitCand :: HoleFitCandidate -> SDoc