GHC.Tc.Gen.Do

Description

Expand Do block statements into (>>=), (>>) and lets After renaming but right ebefore type checking

Documentation

expandDoStmts :: HsDoFlavour -> [ExprLStmt GhcRn] -> TcM (LHsExpr GhcRn)

Expand the `do`-statments into expressions right after renaming so that they can be typechecked. See Note [Expanding HsDo with XXExprGhcRn] below for HsDo specific commentary and Note [Handling overloaded and rebindable constructs] for high level commentary