GHC.Tc.Gen.Foreign

Description

Typechecking foreign declarations

A foreign declaration is used to either give an externally implemented function a Haskell type (and calling interface) or give a Haskell function an external calling interface. Either way, the range of argument and result types these functions can accommodate is restricted to what the outside world understands (read C), and this module checks to see if a foreign declaration has got a legal type.

Documentation

tcForeignImports :: [LForeignDecl GhcRn] -> TcM ([Id], [LForeignDecl GhcTc], Bag GlobalRdrElt)

tcForeignExports :: [LForeignDecl GhcRn] -> TcM (LHsBinds GhcTc, [LForeignDecl GhcTc], Bag GlobalRdrElt)

isForeignImport :: forall name. UnXRec name => LForeignDecl name -> Bool

isForeignExport :: forall name. UnXRec name => LForeignDecl name -> Bool

tcFImport :: LForeignDecl GhcRn -> TcM (Id, LForeignDecl GhcTc, Bag GlobalRdrElt)

tcFExport :: ForeignDecl GhcRn -> TcM (LHsBind GhcTc, ForeignDecl GhcTc, Bag GlobalRdrElt)

tcForeignImports' :: [LForeignDecl GhcRn] -> TcM ([Id], [LForeignDecl GhcTc], Bag GlobalRdrElt)

tcCheckFIType :: [Scaled Type] -> Type -> ForeignImport -> TcM ForeignImport

checkCTarget :: CCallTarget -> TcM ()

checkForeignArgs :: (Type -> Validity) -> [Scaled Type] -> TcM ()

checkForeignRes :: Bool -> Bool -> (Type -> Validity) -> Type -> TcM ()

Check that the type has the form (IO t) or (t) , and that t satisfies the given predicate. When calling this function, any newtype wrappers (should) have been already dealt with by normaliseFfiType.

We also check that the Safe Haskell condition of FFI imports having results in the IO monad holds.

normaliseFfiType :: Type -> TcM (Coercion, Type, Bag GlobalRdrElt)

nonIOok :: Bool

mustBeIO :: Bool

checkSafe :: Bool

noCheckSafe :: Bool

tcForeignExports' :: [LForeignDecl GhcRn] -> TcM (LHsBinds GhcTc, [LForeignDecl GhcTc], Bag GlobalRdrElt)

tcCheckFEType :: Type -> ForeignExport -> TcM ForeignExport