GHC.Tc.Gen.Head

Documentation

data HsExprArg (p :: TcPass) where

Constructors

EValArg

Fields

:: forall (p :: TcPass). { ea_ctxt :: AppCtxt

, ea_arg_ty :: !(XEVAType p)

, ea_arg :: LHsExpr (GhcPass (XPass p))

} -> HsExprArg p

EValArgQL

Fields

:: { eaql_ctxt :: AppCtxt

, eaql_arg_ty :: Scaled TcSigmaType

, eaql_larg :: LHsExpr GhcRn

, eaql_tc_fun :: (HsExpr GhcTc, AppCtxt)

, eaql_fun_ue :: UsageEnv

, eaql_args :: [HsExprArg 'TcpInst]

, eaql_wanted :: WantedConstraints

, eaql_encl :: Bool

, eaql_res_rho :: TcRhoType

} -> HsExprArg 'TcpInst

ETypeArg

Fields

:: forall (p :: TcPass). { ea_ctxt :: AppCtxt

, ea_hs_ty :: LHsWcType GhcRn

, ea_ty_arg :: !(XETAType p)

} -> HsExprArg p

EPrag :: forall (p :: TcPass). AppCtxt -> HsPragE (GhcPass (XPass p)) -> HsExprArg p

EWrap :: forall (p :: TcPass). EWrap -> HsExprArg p

data TcPass

Constructors

TcpRn

TcpInst

TcpTc

data QLFlag

Constructors

DoQL

NoQL

data EWrap

Constructors

EPar AppCtxt

EExpand HsThingRn

EHsWrap HsWrapper

data AppCtxt

Constructors

VAExpansion HsThingRn SrcSpan SrcSpan

VACall (HsExpr GhcRn) Int SrcSpan

appCtxtLoc :: AppCtxt -> SrcSpan

insideExpansion :: AppCtxt -> Bool

splitHsApps :: HsExpr GhcRn -> TcM ((HsExpr GhcRn, AppCtxt), [HsExprArg 'TcpRn])

rebuildHsApps

Arguments

:: (HsExpr GhcTc, AppCtxt)
the function being applied

-> [HsExprArg 'TcpTc]
the arguments to the function

-> HsExpr GhcTc

Rebuild an application: takes a type-checked application head expression together with arguments in the form of typechecked HsExprArgs and returns a typechecked application of the head to the arguments.

This performs a representation-polymorphism check to ensure that representation-polymorphic unlifted newtypes have been eta-expanded.

See Note [Eta-expanding rep-poly unlifted newtypes].

addArgWrap :: forall (p :: TcPass). HsWrapper -> [HsExprArg p] -> [HsExprArg p]

isHsValArg :: forall (id :: TcPass). HsExprArg id -> Bool

leadingValArgs :: [HsExprArg 'TcpRn] -> [LHsExpr GhcRn]

isVisibleArg :: forall (id :: TcPass). HsExprArg id -> Bool

tcInferAppHead :: (HsExpr GhcRn, AppCtxt) -> TcM (HsExpr GhcTc, TcSigmaType)

tcInferAppHead_maybe :: HsExpr GhcRn -> TcM (Maybe (HsExpr GhcTc, TcSigmaType))

tcInferId :: Name -> TcM (HsExpr GhcTc, TcSigmaType)

tcCheckId :: Name -> ExpRhoType -> TcM (HsExpr GhcTc)

obviousSig :: HsExpr GhcRn -> Maybe (LHsSigWcType GhcRn)

tyConOf :: FamInstEnvs -> TcSigmaType -> Maybe TyCon

tyConOfET :: FamInstEnvs -> ExpRhoType -> Maybe TyCon

fieldNotInType :: RecSelParent -> RdrName -> TcRnMessage

nonBidirectionalErr :: Name -> TcRnMessage

pprArgInst :: HsExprArg 'TcpInst -> SDoc

addHeadCtxt :: AppCtxt -> TcM a -> TcM a

addExprCtxt :: HsExpr GhcRn -> TcRn a -> TcRn a

addStmtCtxt :: ExprStmt GhcRn -> TcRn a -> TcRn a

addFunResCtxt :: forall (p :: TcPass) a. HsExpr GhcTc -> [HsExprArg p] -> TcType -> ExpRhoType -> TcM a -> TcM a