GHC.Tc.Solver.Flatten

Documentation

data FlattenMode

Constructors

FM_FlattenAll

FM_SubstOnly

flatten :: FlattenMode -> CtEvidence -> TcType -> TcS (Xi, TcCoercion)

See Note [Flattening]. If (xi, co) <- flatten mode ev ty, then co :: xi ~r ty where r is the role in ev. If mode is FM_FlattenAll, then xi is almost function-free (Note [Almost function-free] in GHC.Tc.Types).

flattenKind :: CtLoc -> CtFlavour -> TcType -> TcS (Xi, TcCoercionN)

flattenArgsNom :: CtEvidence -> TyCon -> [TcType] -> TcS ([Xi], [TcCoercion], TcCoercionN)

rewriteTyVar :: TcTyVar -> TcS TcType

flattenType :: CtLoc -> TcType -> TcS TcType

Flatten a type w.r.t. nominal equality. This is useful to rewrite a type w.r.t. any givens. It does not do type-family reduction. This will never emit new constraints. Call this when the inert set contains only givens.

unflattenWanteds :: Cts -> Cts -> TcS Cts