GHC.Tc.Solver.Monad

Description

Monadic definitions for the constraint solver

Documentation

data TcS a

runTcS :: TcS a -> TcM (a, EvBindMap)

runTcSEarlyAbort :: TcS a -> TcM a

This variant of runTcS will immediately fail upon encountering an insoluble ct. See Note [Speeding up valid hole-fits]. Its one usage site does not need the ev_binds, so we do not return them.

runTcSWithEvBinds :: EvBindsVar -> TcS a -> TcM a

runTcSInerts :: InertSet -> TcS a -> TcM (a, InertSet)

A variant of runTcS that takes and returns an InertSet for later resumption of the TcS session.

failTcS :: TcRnMessage -> TcS a

warnTcS :: TcRnMessage -> TcS ()

addErrTcS :: TcRnMessage -> TcS ()

wrapTcS :: TcM a -> TcS a

ctLocWarnTcS :: CtLoc -> TcRnMessage -> TcS ()

Emit a warning within the TcS monad at the location given by the CtLoc.

runTcSEqualities :: TcS a -> TcM a

This can deal only with equality constraints.

nestTcS :: TcS a -> TcS a

nestImplicTcS :: EvBindsVar -> TcLevel -> TcS a -> TcS a

setEvBindsTcS :: EvBindsVar -> TcS a -> TcS a

emitImplicationTcS :: TcLevel -> SkolemInfoAnon -> [TcTyVar] -> [EvVar] -> Cts -> TcS TcEvBinds

emitTvImplicationTcS :: TcLevel -> SkolemInfoAnon -> [TcTyVar] -> Cts -> TcS ()

emitFunDepWanteds :: CtEvidence -> [FunDepEqn (CtLoc, RewriterSet)] -> TcS Bool

selectNextWorkItem :: TcS (Maybe Ct)

getWorkList :: TcS WorkList

updWorkListTcS :: (WorkList -> WorkList) -> TcS ()

pushLevelNoWorkList :: SDoc -> TcS a -> TcS (TcLevel, a)

runTcPluginTcS :: TcPluginM a -> TcS a

recordUsedGREs :: Bag GlobalRdrElt -> TcS ()

matchGlobalInst :: DynFlags -> Bool -> Class -> [Type] -> CtLoc -> TcS ClsInstResult

data ClsInstResult

Constructors

NoInstance

OneInst

Fields

cir_new_theta :: [TcPredType]

cir_mk_ev :: [EvExpr] -> EvTerm

cir_canonical :: CanonicalEvidence

cir_what :: InstanceWhat

NotSure

data QCInst

Constructors

QCI

Fields

qci_ev :: CtEvidence

qci_tvs :: [TcTyVar]

qci_pred :: TcPredType

qci_pend_sc :: ExpansionFuel

data StopOrContinue a

Constructors

StartAgain Ct

ContinueWith !a

Stop CtEvidence SDoc

continueWith :: a -> TcS (StopOrContinue a)

stopWith :: CtEvidence -> String -> TcS (StopOrContinue a)

startAgainWith :: Ct -> TcS (StopOrContinue a)

newtype SolverStage a

Constructors

Stage

Fields

runSolverStage :: TcS (StopOrContinue a)

simpleStage :: TcS a -> SolverStage a

stopWithStage :: CtEvidence -> String -> SolverStage a

nopStage :: a -> SolverStage a

panicTcS :: SDoc -> TcS a

traceTcS :: String -> SDoc -> TcS ()

tryEarlyAbortTcS :: TcS ()

traceFireTcS :: CtEvidence -> SDoc -> TcS ()

bumpStepCountTcS :: TcS ()

csTraceTcS :: SDoc -> TcS ()

wrapErrTcS :: TcM a -> TcS a

wrapWarnTcS :: TcM a -> TcS a

resetUnificationFlag :: TcS Bool

setUnificationFlag :: TcLevel -> TcS ()

data MaybeNew

Constructors

Fresh CtEvidence

Cached EvExpr

freshGoals :: [MaybeNew] -> [CtEvidence]

isFresh :: MaybeNew -> Bool

getEvExpr :: MaybeNew -> EvExpr

data CanonicalEvidence

CanonicalEvidence says whether a piece of evidence has a singleton type; For example, given (d1 :: C Int), will any other (d2 :: C Int) do equally well? See Note [Coherence and specialisation: overview] above, and Note [Desugaring non-canonical evidence] in GHC.HsToCore.Binds

Constructors

EvCanonical

EvNonCanonical

newTcEvBinds :: TcS EvBindsVar

newNoTcEvBinds :: TcS EvBindsVar

newWantedEq :: CtLoc -> RewriterSet -> Role -> TcType -> TcType -> TcS (CtEvidence, Coercion)

Create a new Wanted constraint holding a coercion hole for an equality between the two types at the given Role.

emitNewWantedEq :: CtLoc -> RewriterSet -> Role -> TcType -> TcType -> TcS Coercion

Emit a new Wanted equality into the work-list

newWanted :: CtLoc -> RewriterSet -> PredType -> TcS MaybeNew

Create a new Wanted constraint, potentially looking up non-equality constraints in the cache instead of creating a new one from scratch.

Deals with both equality and non-equality constraints.

newWantedNC :: CtLoc -> RewriterSet -> PredType -> TcS CtEvidence

Create a new Wanted constraint.

Deals with both equality and non-equality constraints.

Does not attempt to re-use non-equality constraints that already exist in the inert set.

newWantedEvVarNC :: CtLoc -> RewriterSet -> TcPredType -> TcS CtEvidence

Create a new Wanted constraint holding an evidence variable.

Don't use this for equality constraints: use newWantedEq instead.

newBoundEvVarId :: TcPredType -> EvTerm -> TcS EvVar

Make a new Id of the given type, bound (in the monad's EvBinds) to the given term

unifyTyVar :: TcTyVar -> TcType -> TcS ()

reportUnifications :: TcS a -> TcS (Int, a)

setEvBind :: EvBind -> TcS ()

setWantedEq :: HasDebugCallStack => TcEvDest -> Coercion -> TcS ()

Equalities only

setWantedEvTerm :: TcEvDest -> CanonicalEvidence -> EvTerm -> TcS ()

Good for both equalities and non-equalities

setEvBindIfWanted :: CtEvidence -> CanonicalEvidence -> EvTerm -> TcS ()

newEvVar :: TcPredType -> TcS EvVar

newGivenEvVar :: CtLoc -> (TcPredType, EvTerm) -> TcS CtEvidence

emitNewGivens :: CtLoc -> [(Role, TcCoercion)] -> TcS ()

checkReductionDepth

Arguments

:: CtLoc

-> TcType
type being reduced

-> TcS ()

Checks if the depth of the given location is too much. Fails if it's too big, with an appropriate error message.

getSolvedDicts :: TcS (DictMap DictCt)

setSolvedDicts :: DictMap DictCt -> TcS ()

getInstEnvs :: TcS InstEnvs

getFamInstEnvs :: TcS (FamInstEnv, FamInstEnv)

getTopEnv :: TcS HscEnv

getGblEnv :: TcS TcGblEnv

getLclEnv :: TcS TcLclEnv

setSrcSpan :: RealSrcSpan -> TcS a -> TcS a

getTcEvBindsVar :: TcS EvBindsVar

getTcLevel :: TcS TcLevel

getTcEvTyCoVars :: EvBindsVar -> TcS TyCoVarSet

getTcEvBindsMap :: EvBindsVar -> TcS EvBindMap

setTcEvBindsMap :: EvBindsVar -> EvBindMap -> TcS ()

tcLookupClass :: Name -> TcS Class

tcLookupId :: Name -> TcS Id

tcLookupTyCon :: Name -> TcS TyCon

updInertSet :: (InertSet -> InertSet) -> TcS ()

updInertCans :: (InertCans -> InertCans) -> TcS ()

getHasGivenEqs :: TcLevel -> TcS (HasGivenEqs, InertIrreds)

setInertCans :: InertCans -> TcS ()

getInertEqs :: TcS InertEqs

getInertCans :: TcS InertCans

getInertGivens :: TcS [Ct]

getInertInsols :: TcS Cts

Retrieves all insoluble constraints from the inert set, specifically including Given constraints.

This consists of:

insoluble equalities, such as Int ~# Bool;

constraints that are top-level custom type errors, of the form TypeError msg, but not constraints such as Eq (TypeError msg) in which the type error is nested;

unsatisfiable constraints, of the form Unsatisfiable msg.

The inclusion of Givens is important for pattern match warnings, as we want to consider a pattern match that introduces insoluble Givens to be redundant (see Note [Pattern match warnings with insoluble Givens] in GHC.Tc.Solver).

getInnermostGivenEqLevel :: TcS TcLevel

getInertSet :: TcS InertSet

setInertSet :: InertSet -> TcS ()

getUnsolvedInerts :: TcS (Bag Implication, Cts)

removeInertCts :: [Ct] -> InertCans -> InertCans

Remove inert constraints from the InertCans, for use when a typechecker plugin wishes to discard a given.

getPendingGivenScs :: TcS [Ct]

insertFunEq :: FunEqMap a -> TyCon -> [Type] -> a -> FunEqMap a

addInertForAll :: QCInst -> TcS ()

emitWorkNC :: [CtEvidence] -> TcS ()

emitWork :: Cts -> TcS ()

lookupInertDict :: InertCans -> CtLoc -> Class -> [Type] -> Maybe DictCt

Look up a dictionary inert.

kickOutAfterUnification :: [TcTyVar] -> TcS ()

kickOutRewritable :: KickOutSpec -> CtFlavourRole -> TcS ()

addInertSafehask :: InertCans -> DictCt -> InertCans

insertSafeOverlapFailureTcS :: InstanceWhat -> DictCt -> TcS ()

updInertSafehask :: (DictMap DictCt -> DictMap DictCt) -> TcS ()

getSafeOverlapFailures :: TcS (Bag DictCt)

updSolvedDicts :: InstanceWhat -> DictCt -> TcS ()

lookupSolvedDict :: InertSet -> CtLoc -> Class -> [Type] -> Maybe CtEvidence

Look up a solved inert.

foldIrreds :: (IrredCt -> b -> b) -> InertIrreds -> b -> b

lookupFamAppInert :: (CtFlavourRole -> Bool) -> TyCon -> [Type] -> TcS (Maybe (Reduction, CtFlavourRole))

Looks up a family application in the inerts.

lookupFamAppCache :: TyCon -> [Type] -> TcS (Maybe Reduction)

extendFamAppCache :: TyCon -> [Type] -> Reduction -> TcS ()

pprKicked :: Int -> SDoc

instDFunType :: DFunId -> [DFunInstType] -> TcS ([TcType], TcThetaType)

wrapUnifierTcS :: CtEvidence -> Role -> (UnifyEnv -> TcM a) -> TcS (a, Bag Ct, [TcTyVar])

unifyFunDeps :: CtEvidence -> Role -> (UnifyEnv -> TcM ()) -> TcS Bool

uPairsTcM :: UnifyEnv -> [TypeEqn] -> TcM ()

unifyForAllBody :: CtEvidence -> Role -> (UnifyEnv -> TcM a) -> TcS (a, Cts)

newFlexiTcSTy :: Kind -> TcS TcType

instFlexiX :: Subst -> [TKVar] -> TcS Subst

cloneMetaTyVar :: TcTyVar -> TcS TcTyVar

tcInstSkolTyVarsX :: SkolemInfo -> Subst -> [TyVar] -> TcS (Subst, [TcTyVar])

data TcLevel

isFilledMetaTyVar_maybe :: TcTyVar -> TcS (Maybe Type)

isFilledMetaTyVar :: TcTyVar -> TcS Bool

isUnfilledMetaTyVar :: TcTyVar -> TcS Bool

zonkTyCoVarsAndFV :: TcTyCoVarSet -> TcS TcTyCoVarSet

zonkTcType :: TcType -> TcS TcType

zonkTcTypes :: [TcType] -> TcS [TcType]

zonkTcTyVar :: TcTyVar -> TcS TcType

zonkCo :: Coercion -> TcS Coercion

zonkTyCoVarsAndFVList :: [TcTyCoVar] -> TcS [TcTyCoVar]

zonkSimples :: Cts -> TcS Cts

zonkWC :: WantedConstraints -> TcS WantedConstraints

zonkTyCoVarKind :: TcTyCoVar -> TcS TcTyCoVar

newTcRef :: a -> TcS (TcRef a)

readTcRef :: TcRef a -> TcS a

writeTcRef :: TcRef a -> a -> TcS ()

updTcRef :: TcRef a -> (a -> a) -> TcS ()

getDefaultInfo :: TcS (DefaultEnv, Bool)

getDynFlags :: HasDynFlags m => m DynFlags

getGlobalRdrEnvTcS :: TcS GlobalRdrEnv

matchFam :: TyCon -> [Type] -> TcS (Maybe ReductionN)

matchFamTcM :: TyCon -> [Type] -> TcM (Maybe ReductionN)

checkWellStagedDFun :: CtLoc -> InstanceWhat -> PredType -> TcS ()

pprEq :: TcType -> TcType -> SDoc

checkTypeEq :: CtEvidence -> EqRel -> CanEqLHS -> TcType -> TcS (PuResult () Reduction)

checkTouchableTyVarEq :: CtEvidence -> TcTyVar -> TcType -> TcS (PuResult () Reduction)