GHC.Tc.Types.Evidence

HsWrapper

data HsWrapper

Constructors

WpHole

WpCompose HsWrapper HsWrapper

WpFun HsWrapper HsWrapper (Scaled TcTypeFRR)

WpCast TcCoercionR

WpEvLam EvVar

WpEvApp EvTerm

WpTyLam TyVar

WpTyApp KindOrType

WpLet TcEvBinds

WpMultCoercion Coercion

(<.>) :: HsWrapper -> HsWrapper -> HsWrapper

mkWpTyApps :: [Type] -> HsWrapper

mkWpEvApps :: [EvTerm] -> HsWrapper

mkWpEvVarApps :: [EvVar] -> HsWrapper

mkWpTyLams :: [TyVar] -> HsWrapper

mkWpEvLams :: [Var] -> HsWrapper

mkWpLet :: TcEvBinds -> HsWrapper

mkWpFun

Arguments

:: HsWrapper

-> HsWrapper

-> Scaled TcTypeFRR
the "from" type of the first wrapper MUST have a fixed RuntimeRep

-> TcType
Either "from" type or "to" type of the second wrapper (used only when the second wrapper is the identity)

-> HsWrapper

Smart constructor to create a WpFun HsWrapper.

PRECONDITION: the "from" type of the first wrapper must have a syntactically fixed RuntimeRep (see Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete).

mkWpCastN :: TcCoercionN -> HsWrapper

mkWpCastR :: TcCoercionR -> HsWrapper

mkWpEta :: [Id] -> HsWrapper -> HsWrapper

collectHsWrapBinders :: HsWrapper -> ([Var], HsWrapper)

idHsWrapper :: HsWrapper

isIdHsWrapper :: HsWrapper -> Bool

pprHsWrapper :: HsWrapper -> (Bool -> SDoc) -> SDoc

hsWrapDictBinders :: HsWrapper -> Bag DictId

Identifies the lambda-bound dictionaries of an HsWrapper. This is used (only) to allow the pattern-match overlap checker to know what Given dictionaries are in scope.

We specifically do not collect dictionaries bound in a WpLet. These are either superclasses of lambda-bound ones, or (extremely numerous) results of binding Wanted dictionaries. We definitely don't want all those cluttering up the Given dictionaries for pattern-match overlap checking!

Evidence bindings

data TcEvBinds

Constructors

TcEvBinds EvBindsVar

EvBinds (Bag EvBind)

data EvBindsVar

Constructors

EvBindsVar

Fields

ebv_uniq :: Unique

ebv_binds :: IORef EvBindMap

ebv_tcvs :: IORef CoVarSet

CoEvBindsVar

Fields

ebv_uniq :: Unique

ebv_tcvs :: IORef CoVarSet

newtype EvBindMap

Constructors

EvBindMap

Fields

ev_bind_varenv :: DVarEnv EvBind

emptyEvBindMap :: EvBindMap

extendEvBinds :: EvBindMap -> EvBind -> EvBindMap

lookupEvBind :: EvBindMap -> EvVar -> Maybe EvBind

evBindMapBinds :: EvBindMap -> Bag EvBind

foldEvBindMap :: (EvBind -> a -> a) -> a -> EvBindMap -> a

nonDetStrictFoldEvBindMap :: (EvBind -> a -> a) -> a -> EvBindMap -> a

filterEvBindMap :: (EvBind -> Bool) -> EvBindMap -> EvBindMap

isEmptyEvBindMap :: EvBindMap -> Bool

evBindMapToVarSet :: EvBindMap -> VarSet

varSetMinusEvBindMap :: VarSet -> EvBindMap -> VarSet

data EvBind

Constructors

EvBind

Fields

eb_lhs :: EvVar

eb_rhs :: EvTerm

eb_is_given :: Bool

emptyTcEvBinds :: TcEvBinds

isEmptyTcEvBinds :: TcEvBinds -> Bool

mkGivenEvBind :: EvVar -> EvTerm -> EvBind

mkWantedEvBind :: EvVar -> EvTerm -> EvBind

evBindVar :: EvBind -> EvVar

isCoEvBindsVar :: EvBindsVar -> Bool

EvTerm (already a CoreExpr)

data EvTerm

Constructors

EvExpr EvExpr

EvTypeable Type EvTypeable

EvFun

Fields

et_tvs :: [TyVar]

et_given :: [EvVar]

et_binds :: TcEvBinds

et_body :: EvVar

type EvExpr = CoreExpr

evId :: EvId -> EvExpr

Any sort of evidence Id, including coercions

evCoercion :: TcCoercion -> EvTerm

evCast :: EvExpr -> TcCoercion -> EvTerm

d |> co

evDFunApp :: DFunId -> [Type] -> [EvExpr] -> EvTerm

evDataConApp :: DataCon -> [Type] -> [EvExpr] -> EvTerm

evSelector :: Id -> [Type] -> [EvExpr] -> EvExpr

mkEvCast :: EvExpr -> TcCoercion -> EvTerm

evVarsOfTerm :: EvTerm -> VarSet

mkEvScSelectors :: Class -> [TcType] -> [(TcPredType, EvExpr)]

evTypeable :: Type -> EvTypeable -> EvTerm

findNeededEvVars :: EvBindMap -> VarSet -> VarSet

evTermCoercion :: EvTerm -> TcCoercion

evTermCoercion_maybe :: EvTerm -> Maybe TcCoercion

data EvCallStack

Evidence for CallStack implicit parameters.

Constructors

EvCsEmpty

EvCsPushCall FastString RealSrcSpan EvExpr
EvCsPushCall origin loc stk represents a call from origin, occurring at loc, in a calling context stk.

data EvTypeable

Instructions on how to make a Typeable dictionary. See Note [Typeable evidence terms]

Constructors

EvTypeableTyCon TyCon [EvTerm]
Dictionary for Typeable T where T is a type constructor with all of its kind variables saturated. The [EvTerm] is Typeable evidence for the applied kinds..

EvTypeableTyApp EvTerm EvTerm
Dictionary for Typeable (s t), given a dictionaries for s and t.

EvTypeableTrFun EvTerm EvTerm EvTerm
Dictionary for Typeable (s % w -> t), given a dictionaries for w, s, and t.

EvTypeableTyLit EvTerm
Dictionary for a type literal, e.g. Typeable "foo" or Typeable 3 The EvTerm is evidence of, e.g., KnownNat 3 (see #10348)

HoleExprRef

data HoleExprRef

Where to store evidence for expression holes See Note [Holes] in GHC.Tc.Types.Constraint

Constructors

HER

Fields

(IORef EvTerm)

where to write the erroring expression

TcType

expected type of that expression

Unique

for debug output only

TcCoercion

type TcCoercion = Coercion

type TcCoercionR = CoercionR

type TcCoercionN = CoercionN

type TcCoercionP = CoercionP

data CoercionHole

A coercion to be filled in by the type-checker. See Note [Coercion holes]

type TcMCoercion = MCoercion

type TcMCoercionN = MCoercionN

type TcMCoercionR = MCoercionR

data Role

See Note [Roles] in GHC.Core.Coercion

Order of constructors matters: the Ord instance coincides with the *super*typing relation on roles.

Constructors

Nominal

Representational

Phantom

data LeftOrRight

Constructors

CLeft

CRight

pickLR :: LeftOrRight -> (a, a) -> a

maybeSymCo :: SwapFlag -> TcCoercion -> TcCoercion

If a SwapFlag is IsSwapped, flip the orientation of a coercion

unwrapIP :: Type -> CoercionR

Create a Coercion that unwraps an implicit-parameter dictionary to expose the underlying value. We expect the Type to have the form `IP sym ty`, and return a Coercion `co :: IP sym ty ~ ty`

wrapIP :: Type -> CoercionR

Create a Coercion that wraps a value in an implicit-parameter dictionary. See unwrapIP.

QuoteWrapper

data QuoteWrapper

Constructors

QuoteWrapper EvVar Type

applyQuoteWrapper :: QuoteWrapper -> HsWrapper

Convert the QuoteWrapper into a normal HsWrapper which can be used to apply its contents.

quoteWrapperTyVarTy :: QuoteWrapper -> Type