GHC.Tc.Types.TcRef

Documentation

type TcRef a = IORef a

Type alias for IORef; the convention is we'll use this for mutable bits of data in the typechecker which are updated during typechecking and returned at the end.

newTcRef :: MonadIO m => a -> m (TcRef a)

readTcRef :: MonadIO m => TcRef a -> m a

writeTcRef :: MonadIO m => TcRef a -> a -> m ()

updTcRef :: MonadIO m => TcRef a -> (a -> a) -> m ()

updTcRefM :: MonadIO m => TcRef a -> (a -> m a) -> m ()