GHC.Tc.Utils.Instantiate

Documentation

topSkolemise :: TcSigmaType -> TcM (HsWrapper, [(Name, TyVar)], [EvVar], TcRhoType)

topInstantiate :: CtOrigin -> TcSigmaType -> TcM (HsWrapper, TcRhoType)

instantiateSigma :: CtOrigin -> [TyVar] -> TcThetaType -> TcSigmaType -> TcM ([TcTyVar], HsWrapper, TcSigmaType)

instCall :: CtOrigin -> [TcType] -> TcThetaType -> TcM HsWrapper

instDFunType :: DFunId -> [DFunInstType] -> TcM ([TcType], TcThetaType)

instStupidTheta :: CtOrigin -> TcThetaType -> TcM ()

instTyVarsWith :: CtOrigin -> [TyVar] -> [TcType] -> TcM TCvSubst

newWanted :: CtOrigin -> Maybe TypeOrKind -> PredType -> TcM CtEvidence

newWanteds :: CtOrigin -> ThetaType -> TcM [CtEvidence]

tcInstType

Arguments

:: ([TyVar] -> TcM (TCvSubst, [TcTyVar]))
How to instantiate the type variables

-> Id
Type to instantiate

-> TcM ([(Name, TcTyVar)], TcThetaType, TcType)
Result (type vars, preds (incl equalities), rho)

tcInstTypeBndrs :: Id -> TcM ([(Name, InvisTVBinder)], TcThetaType, TcType)

tcInstSkolTyVars :: [TyVar] -> TcM (TCvSubst, [TcTyVar])

Given a list of [TyVar], skolemize the type variables, returning a substitution mapping the original tyvars to the skolems, and the list of newly bound skolems.

tcInstSkolTyVarsX :: TCvSubst -> [TyVar] -> TcM (TCvSubst, [TcTyVar])

tcInstSkolTyVarsAt :: TcLevel -> Bool -> TCvSubst -> [TyVar] -> TcM (TCvSubst, [TcTyVar])

tcSkolDFunType :: DFunId -> TcM ([TcTyVar], TcThetaType, TcType)

tcSuperSkolTyVars :: [TyVar] -> (TCvSubst, [TcTyVar])

tcInstSuperSkolTyVarsX :: TCvSubst -> [TyVar] -> TcM (TCvSubst, [TcTyVar])

freshenTyVarBndrs :: [TyVar] -> TcM (TCvSubst, [TyVar])

Give fresh uniques to a bunch of TyVars, but they stay as TyVars, rather than becoming TcTyVars Used in newFamInst, and newClsInst

freshenCoVarBndrsX :: TCvSubst -> [CoVar] -> TcM (TCvSubst, [CoVar])

Give fresh uniques to a bunch of CoVars Used in "GHC.Tc.Instance.Family.newFamInst"

tcInstInvisibleTyBindersN :: Int -> TcKind -> TcM ([TcType], TcKind)

tcInstInvisibleTyBinders :: TcType -> TcKind -> TcM (TcType, TcKind)

Given ty::forall k1 k2. k, instantiate all the invisible forall-binders returning ty kk1 kk2 :: k[kk1k1, kk2k1]

tcInstInvisibleTyBinder :: TCvSubst -> TyBinder -> TcM (TCvSubst, TcType)

Used only in *types*

newOverloadedLit :: HsOverLit GhcRn -> ExpRhoType -> TcM (HsOverLit GhcTc)

mkOverLit :: OverLitVal -> TcM (HsLit GhcTc)

newClsInst :: Maybe OverlapMode -> Name -> [TyVar] -> ThetaType -> Class -> [Type] -> TcM ClsInst

tcGetInsts :: TcM [ClsInst]

tcGetInstEnvs :: TcM InstEnvs

getOverlapFlag :: Maybe OverlapMode -> TcM OverlapFlag

tcExtendLocalInstEnv :: [ClsInst] -> TcM a -> TcM a

instCallConstraints :: CtOrigin -> TcThetaType -> TcM HsWrapper

newMethodFromName

Arguments

:: CtOrigin
why do we need this?

-> Name
name of the method

-> [TcRhoType]
types with which to instantiate the class

-> TcM (HsExpr GhcTc)

Used when Name is the wired-in name for a wired-in class method, so the caller knows its type for sure, which should be of form

forall a. C a => <blah>

newMethodFromName is supposed to instantiate just the outer type variable and constraint

tcSyntaxName

Arguments

:: CtOrigin

-> TcType
Type to instantiate it at

-> (Name, HsExpr GhcRn)
(Standard name, user name)

-> TcM (Name, HsExpr GhcTc)
(Standard name, suitable expression) USED ONLY FOR CmdTop (sigh) *** See Note [CmdSyntaxTable] in GHC.Hs.Expr

tyCoVarsOfWC :: WantedConstraints -> TyCoVarSet

Returns free variables of WantedConstraints as a non-deterministic set. See Note [Deterministic FV] in GHC.Utils.FV.

tyCoVarsOfCt :: Ct -> TcTyCoVarSet

Returns free variables of constraints as a non-deterministic set

tyCoVarsOfCts :: Cts -> TcTyCoVarSet

Returns free variables of a bag of constraints as a non-deterministic set. See Note [Deterministic FV] in GHC.Utils.FV.