GHC.Tc.Utils.TcType

Description

Types used in the typechecker

This module provides the Type interface for front-end parts of the compiler. These parts

treat "source types" as opaque: newtypes, and predicates are meaningful.

look through usage types

Documentation

type TcType = Type

type TcSigmaType = TcType

type TcTypeFRR = TcType

A type which has a syntactically fixed RuntimeRep as per Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete.

type TcSigmaTypeFRR = TcSigmaType

A TcSigmaTypeFRR is a TcSigmaType which has a syntactically fixed RuntimeRep in the sense of Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete.

In particular, this means that:

typePrimRep does not panic,

typeLevity_maybe does not return Nothing.

This property is important in functions such as matchExpectedFunTys, where we want to provide argument types which have a known runtime representation. See Note [Return arguments with a fixed RuntimeRep.

type TcRhoType = TcType

type TcTauType = TcType

type TcPredType = PredType

type TcThetaType = ThetaType

type TcTyVar = Var

Type variable that might be a metavariable

type TcTyVarSet = TyVarSet

type TcDTyVarSet = DTyVarSet

type TcTyCoVarSet = TyCoVarSet

type TcDTyCoVarSet = DTyCoVarSet

type TcKind = Kind

type TcCoVar = CoVar

type TcTyCoVar = Var

type TcTyVarBinder = TyVarBinder

type TcInvisTVBinder = InvisTVBinder

type TcReqTVBinder = ReqTVBinder

type TcTyCon = TyCon

type MonoTcTyCon = TcTyCon

type PolyTcTyCon = TcTyCon

type TcTyConBinder = TyConBinder

type KnotTied ty = ty

A type labeled KnotTied might have knot-tied tycons in it. See Note [Type checking recursive type and class declarations] in GHC.Tc.TyCl

data ExpType

An expected type to check against during type-checking. See Note [ExpType] in GHC.Tc.Utils.TcMType, where you'll also find manipulators.

Constructors

Check TcType

Infer !InferResult

data InferResult

Constructors

IR

Fields

ir_uniq :: Unique

This Unique is for debugging only

ir_lvl :: TcLevel

See Note [TcLevel of ExpType] in GHC.Tc.Utils.TcMType

ir_frr :: Maybe FixedRuntimeRepContext

See Note [FixedRuntimeRep context in ExpType] in GHC.Tc.Utils.TcMType

ir_ref :: IORef (Maybe TcType)

The type that fills in this hole should be a Type, that is, its kind should be TYPE rr for some rr :: RuntimeRep.

Additionally, if the ir_frr field is Just frr_orig then rr must be concrete, in the sense of Note [Concrete types] in GHC.Tc.Utils.Concrete.

type ExpTypeFRR = ExpType

An ExpType which has a fixed RuntimeRep.

For a Check ExpType, the stored TcType must have a fixed RuntimeRep. For an Infer ExpType, the ir_frr field must be of the form Just frr_orig.

type ExpSigmaType = ExpType

type ExpSigmaTypeFRR = ExpTypeFRR

Like TcSigmaTypeFRR, but for an expected type.

See ExpTypeFRR.

type ExpRhoType = ExpType

mkCheckExpType :: TcType -> ExpType

Make an ExpType suitable for checking.

data SyntaxOpType

What to expect for an argument to a rebindable-syntax operator. Quite like Type, but allows for holes to be filled in by tcSyntaxOp. The callback called from tcSyntaxOp gets a list of types; the meaning of these types is determined by a left-to-right depth-first traversal of the SyntaxOpType tree. So if you pass in

SynAny `SynFun` (SynList `SynFun` SynType Int) `SynFun` SynAny

you'll get three types back: one for the first SynAny, the element type of the list, and one for the last SynAny. You don't get anything for the SynType, because you've said positively that it should be an Int, and so it shall be.

You'll also get three multiplicities back: one for each function arrow. See also Note [Linear types] in Multiplicity.

This is defined here to avoid defining it in GHC.Tc.Gen.Expr boot file.

Constructors

SynAny
Any type

SynRho
A rho type, skolemised or instantiated as appropriate

SynList
A list type. You get back the element type of the list

SynFun SyntaxOpType SyntaxOpType infixr 0
A function.

SynType ExpType
A known type.

synKnownType :: TcType -> SyntaxOpType

Like SynType but accepts a regular TcType

mkSynFunTys :: [SyntaxOpType] -> ExpType -> SyntaxOpType

Like mkFunTys but for SyntaxOpType

newtype TcLevel

Constructors

TcLevel Int

topTcLevel :: TcLevel

pushTcLevel :: TcLevel -> TcLevel

isTopTcLevel :: TcLevel -> Bool

strictlyDeeperThan :: TcLevel -> TcLevel -> Bool

deeperThanOrSame :: TcLevel -> TcLevel -> Bool

sameDepthAs :: TcLevel -> TcLevel -> Bool

tcTypeLevel :: TcType -> TcLevel

tcTyVarLevel :: TcTyVar -> TcLevel

maxTcLevel :: TcLevel -> TcLevel -> TcLevel

data TcTyVarDetails

Constructors

SkolemTv SkolemInfo TcLevel Bool

RuntimeUnk

MetaTv

Fields

mtv_info :: MetaInfo

mtv_ref :: IORef MetaDetails

mtv_tclvl :: TcLevel

pprTcTyVarDetails :: TcTyVarDetails -> SDoc

vanillaSkolemTvUnk :: HasCallStack => TcTyVarDetails

data MetaDetails

Constructors

Flexi

Indirect TcType

data MetaInfo

What restrictions are on this metavariable around unification? These are checked in GHC.Tc.Utils.Unify.startSolvingByUnification.

Constructors

TauTv
This MetaTv is an ordinary unification variable A TauTv is always filled in with a tau-type, which never contains any ForAlls.

TyVarTv
A variant of TauTv, except that it should not be unified with a type, only with a type variable See Note [TyVarTv] in GHC.Tc.Utils.TcMType

RuntimeUnkTv
A unification variable used in the GHCi debugger. It is allowed to unify with a polytype, unlike TauTv

CycleBreakerTv

ConcreteTv ConcreteTvOrigin
A unification variable that can only be unified with a concrete type, in the sense of Note [Concrete types] in GHC.Tc.Utils.Concrete. See Note [ConcreteTv] in GHC.Tc.Utils.Concrete. See also Note [The Concrete mechanism] in GHC.Tc.Utils.Concrete for an overview of how this works in context.

skolemSkolInfo :: TcTyVar -> SkolemInfo

isImmutableTyVar :: TyVar -> Bool

isSkolemTyVar :: TcTyVar -> Bool

isMetaTyVar :: TcTyVar -> Bool

isMetaTyVarTy :: TcType -> Bool

isTyVarTy :: Type -> Bool

tcIsTcTyVar :: TcTyVar -> Bool

isTyVarTyVar :: Var -> Bool

isOverlappableTyVar :: TcTyVar -> Bool

isTyConableTyVar :: TcTyVar -> Bool

data ConcreteTvOrigin

What caused us to create a ConcreteTv metavariable? See Note [ConcreteTv] in GHC.Tc.Utils.Concrete.

Constructors

ConcreteFRR FixedRuntimeRepOrigin
A ConcreteTv used to enforce the representation-polymorphism invariants.

See FixedRuntimeRepOrigin for more information.

isConcreteTyVar_maybe :: TcTyVar -> Maybe ConcreteTvOrigin

Is this type variable a concrete type variable, i.e. it is a metavariable with ConcreteTv MetaInfo?

Returns the ConcreteTvOrigin stored in the type variable if so, or Nothing otherwise.

isConcreteTyVar :: TcTyVar -> Bool

Is this type variable a concrete type variable, i.e. it is a metavariable with ConcreteTv MetaInfo?

isConcreteTyVarTy :: TcType -> Bool

Is this type concrete type variable, i.e. a metavariable with ConcreteTv MetaInfo?

isConcreteTyVarTy_maybe :: TcType -> Maybe (TcTyVar, ConcreteTvOrigin)

Is this type a concrete type variable? If so, return the associated TcTyVar and ConcreteTvOrigin.

isAmbiguousTyVar :: TcTyVar -> Bool

isCycleBreakerTyVar :: TcTyVar -> Bool

metaTyVarRef :: TyVar -> IORef MetaDetails

metaTyVarInfo :: TcTyVar -> MetaInfo

isFlexi :: MetaDetails -> Bool

isIndirect :: MetaDetails -> Bool

isRuntimeUnkSkol :: TyVar -> Bool

metaTyVarTcLevel :: TcTyVar -> TcLevel

setMetaTyVarTcLevel :: TcTyVar -> TcLevel -> TcTyVar

metaTyVarTcLevel_maybe :: TcTyVar -> Maybe TcLevel

isTouchableMetaTyVar :: TcLevel -> TcTyVar -> Bool

isPromotableMetaTyVar :: TcTyVar -> Bool

findDupTyVarTvs :: [(Name, TcTyVar)] -> [(Name, Name)]

mkTyVarNamePairs :: [TyVar] -> [(Name, TyVar)]

mkInfSigmaTy :: HasDebugCallStack => [TyCoVar] -> [PredType] -> Type -> Type

Make a sigma ty where all type variables are Inferred. That is, they cannot be used with visible type application.

mkSpecSigmaTy :: HasDebugCallStack => [TyVar] -> [PredType] -> Type -> Type

Make a sigma ty where all type variables are "specified". That is, they can be used with visible type application

mkSigmaTy :: HasDebugCallStack => [ForAllTyBinder] -> [PredType] -> Type -> Type

mkPhiTy :: HasDebugCallStack => [PredType] -> Type -> Type

tcMkPhiTy :: HasDebugCallStack => [PredType] -> Type -> Type

tcMkDFunSigmaTy :: [TyVar] -> ThetaType -> Type -> Type

tcMkDFunPhiTy :: HasDebugCallStack => [PredType] -> Type -> Type

getTyVar :: HasDebugCallStack => Type -> TyVar

Attempts to obtain the type variable underlying a Type, and panics with the given message if this is not a type variable type. See also getTyVar_maybe

getTyVar_maybe :: Type -> Maybe TyVar

Attempts to obtain the type variable underlying a Type

getCastedTyVar_maybe :: Type -> Maybe (TyVar, CoercionN)

If the type is a tyvar, possibly under a cast, returns it, along with the coercion. Thus, the co is :: kind tv ~N kind ty

tcSplitForAllTyVarBinder_maybe :: Type -> Maybe (TyVarBinder, Type)

tcSplitForAllTyVars :: Type -> ([TyVar], Type)

Like tcSplitPiTys, but splits off only named binders, returning just the tyvars.

tcSplitForAllInvisTyVars :: Type -> ([TyVar], Type)

Like tcSplitForAllTyVars, but only splits ForAllTys with Invisible type variable binders.

tcSplitSomeForAllTyVars :: (ForAllTyFlag -> Bool) -> Type -> ([TyVar], Type)

Like tcSplitForAllTyVars, but only splits a ForAllTy if argf_pred argf is True, where argf is the visibility of the ForAllTy's binder and argf_pred is a predicate over visibilities provided as an argument to this function.

tcSplitForAllReqTVBinders :: Type -> ([TcReqTVBinder], Type)

Like tcSplitForAllTyVars, but only splits ForAllTys with Required type variable binders. All split tyvars are annotated with ().

tcSplitForAllInvisTVBinders :: Type -> ([TcInvisTVBinder], Type)

Like tcSplitForAllTyVars, but only splits ForAllTys with Invisible type variable binders. All split tyvars are annotated with their Specificity.

tcSplitPiTys :: Type -> ([PiTyVarBinder], Type)

Splits a forall type into a list of PiTyVarBinders and the inner type. Always succeeds, even if it returns an empty list.

tcSplitPiTy_maybe :: Type -> Maybe (PiTyVarBinder, Type)

Splits a type into a PiTyVarBinder and a body, if possible.

tcSplitForAllTyVarBinders :: Type -> ([TyVarBinder], Type)

Like tcSplitForAllTyVars, but splits off only named binders.

tcSplitPhiTy :: Type -> (ThetaType, Type)

tcSplitPredFunTy_maybe :: Type -> Maybe (PredType, Type)

tcSplitFunTy_maybe :: Type -> Maybe (Scaled Type, Type)

tcSplitFunTys :: Type -> ([Scaled Type], Type)

tcFunArgTy :: Type -> Scaled Type

tcFunResultTy :: Type -> Type

tcFunResultTyN :: HasDebugCallStack => Arity -> Type -> Type

Strips off n *visible* arguments and returns the resulting type

tcSplitFunTysN :: Arity -> TcRhoType -> Either Arity ([Scaled TcSigmaType], TcSigmaType)

Split off exactly the specified number argument types Returns (Left m) if there are m missing arrows in the type (Right (tys,res)) if the type looks like t1 -> ... -> tn -> res

tcSplitTyConApp :: Type -> (TyCon, [Type])

tcSplitTyConApp_maybe :: HasCallStack => Type -> Maybe (TyCon, [Type])

tcSplitTyConApp_maybe splits a type constructor application into its type constructor and applied types.

Differs from splitTyConApp_maybe in that it does *not* split types headed with (=>), as that's not a TyCon in the type-checker.

Note that this may fail (in funTyConAppTy_maybe) in the case of a FunTy with an argument of unknown kind FunTy (e.g. `FunTy (a :: k) Int`, since the kind of a isn't of the form `TYPE rep`. This isn't usually a problem but may be temporarily the cas during canonicalization: see Note [Decomposing FunTy] in GHC.Tc.Solver.Canonical and Note [The Purely Kinded Type Invariant (PKTI)] in GHC.Tc.Gen.HsType, Wrinkle around FunTy

Consequently, you may need to zonk your type before using this function.

tcTyConAppTyCon :: Type -> TyCon

tcTyConAppTyCon_maybe :: Type -> Maybe TyCon

Like tcRepSplitTyConApp_maybe, but only returns the TyCon.

tcTyConAppArgs :: Type -> [Type]

tcSplitAppTy_maybe :: Type -> Maybe (Type, Type)

tcSplitAppTy :: Type -> (Type, Type)

tcSplitAppTys :: Type -> (Type, [Type])

tcSplitAppTyNoView_maybe :: Type -> Maybe (Type, Type)

Just like splitAppTyNoView_maybe, but does not split (c => t) See Note [Decomposing fat arrow c=>t]

tcSplitSigmaTy :: Type -> ([TyVar], ThetaType, Type)

Split a sigma type into its parts. This only splits invisible type variable binders, as these are the only forms of binder that the typechecker will implicitly instantiate.

tcSplitNestedSigmaTys :: Type -> ([TyVar], ThetaType, Type)

Split a sigma type into its parts, going underneath as many arrows and foralls as possible. See Note [tcSplitNestedSigmaTys]

isSigmaTy :: TcType -> Bool

isRhoTy :: TcType -> Bool

isRhoExpTy :: ExpType -> Bool

Like isRhoTy, but also says True for Infer types

isOverloadedTy :: Type -> Bool

isFloatingPrimTy :: Type -> Bool

Is the type inhabited by machine floating-point numbers?

Used to check that we don't use floating-point literal patterns in Core.

See #9238 and Note [Rules for floating-point comparisons] in GHC.Core.Opt.ConstantFold.

isDoubleTy :: Type -> Bool

isFloatTy :: Type -> Bool

isIntTy :: Type -> Bool

isWordTy :: Type -> Bool

isStringTy :: Type -> Bool

Is a type String?

isIntegerTy :: Type -> Bool

isNaturalTy :: Type -> Bool

isBoolTy :: Type -> Bool

isUnitTy :: Type -> Bool

isCharTy :: Type -> Bool

isTauTy :: Type -> Bool

isTauTyCon :: TyCon -> Bool

tcIsTyVarTy :: Type -> Bool

isPredTy :: HasDebugCallStack => Type -> Bool

isTyVarClassPred :: PredType -> Bool

checkValidClsArgs :: Bool -> Class -> [KindOrType] -> Bool

hasTyVarHead :: Type -> Bool

isRigidTy :: TcType -> Bool

eqType :: Type -> Type -> Bool

Type equality on source types. Does not look through newtypes, PredTypes or type families, but it does look through type synonyms. This first checks that the kinds of the types are equal and then checks whether the types are equal, ignoring casts and coercions. (The kind check is a recursive call, but since all kinds have type Type, there is no need to check the types of kinds.) See also Note [Non-trivial definitional equality] in GHC.Core.TyCo.Rep.

eqTypes :: [Type] -> [Type] -> Bool

Type equality on lists of types, looking through type synonyms but not newtypes.

nonDetCmpType :: Type -> Type -> Ordering

nonDetCmpTypes :: [Type] -> [Type] -> Ordering

eqTypeX :: RnEnv2 -> Type -> Type -> Bool

Compare types with respect to a (presumably) non-empty RnEnv2.

pickyEqType :: Type -> Type -> Bool

Like pickyEqTypeVis, but returns a Bool for convenience

tcEqType :: HasDebugCallStack => Type -> Type -> Bool

tcEqType implements typechecker equality It behaves just like eqType, but is implemented differently (for now)

tcEqKind :: HasDebugCallStack => Kind -> Kind -> Bool

tcEqTypeNoKindCheck :: Type -> Type -> Bool

Just like tcEqType, but will return True for types of different kinds as long as their non-coercion structure is identical.

tcEqTypeVis :: Type -> Type -> Bool

Like tcEqType, but returns True if the visible part of the types are equal, even if they are really unequal (in the invisible bits)

tcEqTyConApps :: TyCon -> [Type] -> TyCon -> [Type] -> Bool

Check whether two TyConApps are the same; if the number of arguments are different, just checks the common prefix of arguments.

eqForAllVis :: ForAllTyFlag -> ForAllTyFlag -> Bool

Do these denote the same level of visibility? Required arguments are visible, others are not. So this function equates Specified and Inferred. Used for printing.

eqVarBndrs :: RnEnv2 -> [Var] -> [Var] -> Maybe RnEnv2

deNoteType :: Type -> Type

orphNamesOfType :: Type -> NameSet

orphNamesOfCo :: Coercion -> NameSet

orphNamesOfTypes :: [Type] -> NameSet

orphNamesOfCoCon :: CoAxiom br -> NameSet

getDFunTyKey :: Type -> OccName

evVarPred :: EvVar -> PredType

ambigTkvsOfTy :: TcType -> ([Var], [Var])

Returns the (kind, type) variables in a type that are as-yet-unknown: metavariables and RuntimeUnks

mkMinimalBySCs :: forall a. (a -> PredType) -> [a] -> [a]

transSuperClasses :: PredType -> [PredType]

pickCapturedPreds :: TyVarSet -> TcThetaType -> TcThetaType

immSuperClasses :: Class -> [Type] -> [PredType]

boxEqPred :: EqRel -> Type -> Type -> Maybe (Class, [Type])

isImprovementPred :: PredType -> Bool

Finding type instances

tcTyFamInsts :: Type -> [(TyCon, [Type])]

Finds outermost type-family applications occurring in a type, after expanding synonyms. In the list (F, tys) that is returned we guarantee that tys matches F's arity. For example, given type family F a :: * -> * (arity 1) calling tcTyFamInsts on (Maybe (F Int Bool) will return (F, [Int]), not (F, [Int,Bool])

This is important for its use in deciding termination of type instances (see #11581). E.g. type instance G [Int] = ...(F Int <big type>)... we don't need to take <big type> into account when asking if the calls on the RHS are smaller than the LHS

tcTyFamInstsAndVis :: Type -> [(Bool, TyCon, [Type])]

Like tcTyFamInsts, except that the output records whether the type family and its arguments occur as an invisible argument in some type application. This information is useful because it helps GHC know when to turn on -fprint-explicit-kinds during error reporting so that users can actually see the type family being mentioned.

As an example, consider:

class C a
data T (a :: k)
type family F a :: k
instance C (T @(F Int) (F Bool))

There are two occurrences of the type family F in that C instance, so tcTyFamInstsAndVis (C (T @(F Int) (F Bool))) will return:

[ (True,  F, [Int])
, (False, F, [Bool]) ]

F Int is paired with True since it appears as an invisible argument to C, whereas F Bool is paired with False since it appears an a visible argument to C.

See also Note [Kind arguments in error messages] in GHC.Tc.Errors.

tcTyConAppTyFamInstsAndVis :: TyCon -> [Type] -> [(Bool, TyCon, [Type])]

In an application of a TyCon to some arguments, find the outermost occurrences of type family applications within the arguments. This function will not consider the TyCon itself when checking for type family applications.

See tcTyFamInstsAndVis for more details on how this works (as this function is called inside of tcTyFamInstsAndVis).

isTyFamFree :: Type -> Bool

Check that a type does not contain any type family applications.

Finding "exact" (non-dead) type variables

exactTyCoVarsOfType :: Type -> TyCoVarSet

exactTyCoVarsOfTypes :: [Type] -> TyCoVarSet

anyRewritableTyVar :: EqRel -> (EqRel -> TcTyVar -> Bool) -> TcType -> Bool

anyRewritableTyFamApp :: EqRel -> (EqRel -> TyCon -> [TcType] -> Bool) -> TcType -> Bool

data IllegalForeignTypeReason

Reason why a type in an FFI signature is invalid

Constructors

TypeCannotBeMarshaled !Type TypeCannotBeMarshaledReason

ForeignDynNotPtr

Fields

!Type

Expected type

!Type

Actual type

SafeHaskellMustBeInIO

IOResultExpected

UnexpectedNestedForall

LinearTypesNotAllowed

OneArgExpected

AtLeastOneArgExpected

data TypeCannotBeMarshaledReason

Reason why a type cannot be marshalled through the FFI.

Constructors

NotADataType

NewtypeDataConNotInScope !(Maybe TyCon)

UnliftedFFITypesNeeded

NotABoxedMarshalableTyCon

ForeignLabelNotAPtr

NotSimpleUnliftedType

NotBoxedKindAny

isFFIArgumentTy :: DynFlags -> Safety -> Type -> Validity' IllegalForeignTypeReason

isFFIImportResultTy :: DynFlags -> Type -> Validity' IllegalForeignTypeReason

isFFIExportResultTy :: Type -> Validity' IllegalForeignTypeReason

isFFIExternalTy :: Type -> Validity' IllegalForeignTypeReason

isFFIDynTy :: Type -> Type -> Validity' IllegalForeignTypeReason

isFFIPrimArgumentTy :: DynFlags -> Type -> Validity' IllegalForeignTypeReason

isFFIPrimResultTy :: DynFlags -> Type -> Validity' IllegalForeignTypeReason

isFFILabelTy :: Type -> Validity' IllegalForeignTypeReason

isFunPtrTy :: Type -> Bool

tcSplitIOType_maybe :: Type -> Maybe (TyCon, Type)

data PatersonSize

The Paterson size of a given type, in the sense of Note [Paterson conditions] in GHC.Tc.Validity

after expanding synonyms,

ignoring coercions (as they are not user written).

Constructors

PS_TyFam TyCon
The type mentions a type family, so the size could be anything.

PS_Vanilla
The type does not mention a type family.

Fields

ps_tvs :: [TyVar]

free tyvars, including repetitions;

ps_size :: Int

number of type constructors and variables

data PatersonSizeFailure

Why was the LHS PatersonSize not strictly smaller than the RHS PatersonSize?

See Note [Paterson conditions] in GHC.Tc.Validity.

Constructors

PSF_TyFam TyCon
Either side contains a type family.

PSF_Size
The size of the LHS is not strictly less than the size of the RHS.

PSF_TyVar
These type variables appear more often in the LHS than in the RHS.

Fields

[TyVar]

no duplicates in this list

ltPatersonSize :: PatersonSize -> PatersonSize -> Maybe PatersonSizeFailure

ltPatersonSize ps1 ps2 returns:

Nothing iff ps1 is definitely strictly smaller than ps2,

Just ps_fail otherwise; ps_fail says what went wrong.

pSizeZero :: PatersonSize

pSizeOne :: PatersonSize

pSizeType :: Type -> PatersonSize

pSizeTypeX :: VarSet -> Type -> PatersonSize

pSizeTypes :: [Type] -> PatersonSize

pSizeClassPred :: Class -> [Type] -> PatersonSize

pSizeClassPredX :: VarSet -> Class -> [Type] -> PatersonSize

pSizeTyConApp :: TyCon -> [Type] -> PatersonSize

noMoreTyVars :: [TyVar] -> [TyVar] -> [TyVar]

allDistinctTyVars :: TyVarSet -> [KindOrType] -> Bool

type TypeSize = IntWithInf

sizeType :: Type -> TypeSize

sizeTypes :: [Type] -> TypeSize

scopedSort :: [TyCoVar] -> [TyCoVar]

Do a topological sort on a list of tyvars, so that binders occur before occurrences E.g. given [ a::k, k::*, b::k ] it'll return a well-scoped list [ k::*, a::k, b::k ]

This is a deterministic sorting operation (that is, doesn't depend on Uniques).

It is also meant to be stable: that is, variables should not be reordered unnecessarily. This is specified in Note [ScopedSort] See also Note [Ordering of implicit variables] in GHC.Rename.HsType

isTerminatingClass :: Class -> Bool

When this says True, ignore this class constraint during a termination check See (PS1) in Note [The PatersonSize of a type]

isStuckTypeFamily :: TyCon -> Bool

type Kind = Type

The key type representing kinds in the compiler.

liftedTypeKind :: Type

constraintKind :: Kind

isLiftedTypeKind :: Kind -> Bool

Returns True if the argument is (lifted) Type or Constraint See Note [TYPE and CONSTRAINT] in GHC.Builtin.Types.Prim

isUnliftedTypeKind :: Kind -> Bool

Returns True if the kind classifies unlifted types (like 'Int#') and False otherwise. Note that this returns False for representation-polymorphic kinds, which may be specialized to a kind that classifies unlifted types.

isTYPEorCONSTRAINT :: Kind -> Bool

Does this classify a type allowed to have values? Responds True to things like *, TYPE Lifted, TYPE IntRep, TYPE v, Constraint.

True of a kind `TYPE _` or `CONSTRAINT _`

data Type

type PredType = Type

A type of the form p of constraint kind represents a value whose type is the Haskell predicate p, where a predicate is what occurs before the => in a Haskell type.

We use PredType as documentation to mark those types that we guarantee to have this kind.

It can be expanded into its representation, but:

The type checker must treat it as opaque

The rest of the compiler treats it as transparent

Consider these examples:

f :: (Eq a) => a -> Int
g :: (?x :: Int -> Int) => a -> Int
h :: (r\l) => {r} => {l::Int | r}

Here the Eq a and ?x :: Int -> Int and rl are all called "predicates"

type ThetaType = [PredType]

A collection of PredTypes

data PiTyBinder

A PiTyBinder represents an argument to a function. PiTyBinders can be dependent (Named) or nondependent (Anon). They may also be visible or not. See Note [PiTyBinders]

data ForAllTyFlag

ForAllTyFlag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, ForAllTyBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible Specificity

Required

Bundled Patterns

pattern Specified :: ForAllTyFlag

pattern Inferred :: ForAllTyFlag

data FunTyFlag

The non-dependent version of ForAllTyFlag. See Note [FunTyFlag] Appears here partly so that it's together with its friends ForAllTyFlag and ForallVisFlag, but also because it is used in IfaceType, rather early in the compilation chain

Constructors

FTF_T_T

FTF_T_C

FTF_C_T

FTF_C_C

mkForAllTy :: ForAllTyBinder -> Type -> Type

Like mkTyCoForAllTy, but does not check the occurrence of the binder See Note [Unused coercion variable in ForAllTy]

mkForAllTys :: [ForAllTyBinder] -> Type -> Type

Wraps foralls over the type using the provided TyCoVars from left to right

mkInvisForAllTys :: [InvisTVBinder] -> Type -> Type

Wraps foralls over the type using the provided InvisTVBinders from left to right

mkTyCoInvForAllTys :: [TyCoVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and Inferred, a common case

mkSpecForAllTys :: [TyVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and Specified, a common case

mkTyCoInvForAllTy :: TyCoVar -> Type -> Type

Make a dependent forall over an Inferred variable

mkInfForAllTy :: TyVar -> Type -> Type

Like mkTyCoInvForAllTy, but tv should be a tyvar

mkInfForAllTys :: [TyVar] -> Type -> Type

Like mkTyCoInvForAllTys, but tvs should be a list of tyvar

mkVisFunTy :: HasDebugCallStack => Mult -> Type -> Type -> Type

mkVisFunTyMany :: HasDebugCallStack => Type -> Type -> Type infixr 3

Make nested arrow types | Special, common, case: Arrow type with mult Many

mkVisFunTysMany :: [Type] -> Type -> Type

mkScaledFunTys :: HasDebugCallStack => [Scaled Type] -> Type -> Type

mkInvisFunTy :: HasDebugCallStack => Type -> Type -> Type infixr 3

mkInvisFunTys :: HasDebugCallStack => [Type] -> Type -> Type

mkTyConApp :: TyCon -> [Type] -> Type

A key function: builds a TyConApp or FunTy as appropriate to its arguments. Applies its arguments to the constructor from left to right.

mkAppTy :: Type -> Type -> Type

Applies a type to another, as in e.g. k a

mkAppTys :: Type -> [Type] -> Type

mkTyConTy :: TyCon -> Type

(mkTyConTy tc) returns (TyConApp tc []) but arranges to share that TyConApp among all calls See Note [Sharing nullary TyConApps] So it's just an alias for tyConNullaryTy!

mkTyVarTy :: TyVar -> Type

mkTyVarTys :: [TyVar] -> [Type]

mkTyCoVarTy :: TyCoVar -> Type

mkTyCoVarTys :: [TyCoVar] -> [Type]

isClassPred :: PredType -> Bool

isEqPrimPred :: PredType -> Bool

isIPLikePred :: Type -> Bool

isEqPred :: PredType -> Bool

isEqPredClass :: Class -> Bool

mkClassPred :: Class -> [Type] -> PredType

tcSplitQuantPredTy :: Type -> ([TyVar], [Type], PredType)

tcSplitDFunTy :: Type -> ([TyVar], [Type], Class, [Type])

tcSplitDFunHead :: Type -> (Class, [Type])

tcSplitMethodTy :: Type -> ([TyVar], PredType, Type)

isRuntimeRepVar :: TyVar -> Bool

Is a tyvar of type RuntimeRep?

isFixedRuntimeRepKind :: HasDebugCallStack => Kind -> Bool

Checks that a kind of the form Type, Constraint or 'TYPE r is concrete. See isConcrete.

Precondition: The type has kind `TYPE blah` or `CONSTRAINT blah`

isVisiblePiTyBinder :: PiTyBinder -> Bool

Does this binder bind a visible argument?

isInvisiblePiTyBinder :: PiTyBinder -> Bool

Does this binder bind an invisible argument?

data Subst

Type & coercion & id substitution

The Subst data type defined in this module contains substitution for tyvar, covar and id. However, operations on IdSubstEnv (mapping from Id to CoreExpr) that require the definition of the Expr data type are defined in GHC.Core.Subst to avoid circular module dependency.

Constructors

Subst InScopeSet IdSubstEnv TvSubstEnv CvSubstEnv

type TvSubstEnv = TyVarEnv Type

A substitution of Types for TyVars and Kinds for KindVars

emptySubst :: Subst

mkEmptySubst :: InScopeSet -> Subst

zipTvSubst :: HasDebugCallStack => [TyVar] -> [Type] -> Subst

Generates the in-scope set for the Subst from the types in the incoming environment. No CoVars or Ids, please!

mkTvSubstPrs :: [(TyVar, Type)] -> Subst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please! The InScopeSet is just a thunk so with a bit of luck it'll never be evaluated

notElemSubst :: Var -> Subst -> Bool

unionSubst :: Subst -> Subst -> Subst

getTvSubstEnv :: Subst -> TvSubstEnv

getSubstInScope :: Subst -> InScopeSet

Find the in-scope set: see Note [The substitution invariant]

extendSubstInScope :: Subst -> Var -> Subst

Add the Var to the in-scope set

extendSubstInScopeList :: Subst -> [Var] -> Subst

Add the Vars to the in-scope set: see also extendInScope

extendSubstInScopeSet :: Subst -> VarSet -> Subst

Add the Vars to the in-scope set: see also extendInScope

extendTvSubstAndInScope :: Subst -> TyVar -> Type -> Subst

lookupTyVar :: Subst -> TyVar -> Maybe Type

extendTCvSubst :: Subst -> TyCoVar -> Type -> Subst

substTyVarBndr :: HasDebugCallStack => Subst -> TyVar -> (Subst, TyVar)

extendTvSubst :: Subst -> TyVar -> Type -> Subst

Add a substitution for a TyVar to the Subst The TyVar *must* be a real TyVar, and not a CoVar You must ensure that the in-scope set is such that Note [The substitution invariant] holds after extending the substitution like this.

isInScope :: Var -> Subst -> Bool

mkSubst :: InScopeSet -> TvSubstEnv -> CvSubstEnv -> IdSubstEnv -> Subst

mkTvSubst :: InScopeSet -> TvSubstEnv -> Subst

Make a TCvSubst with specified tyvar subst and empty covar subst

zipTyEnv :: HasDebugCallStack => [TyVar] -> [Type] -> TvSubstEnv

The InScopeSet is just a thunk so with a bit of luck it'll never be evaluated

zipCoEnv :: HasDebugCallStack => [CoVar] -> [Coercion] -> CvSubstEnv

substTy :: HasDebugCallStack => Subst -> Type -> Type

Substitute within a Type The substitution has to satisfy the invariants described in Note [The substitution invariant].

substTys :: HasDebugCallStack => Subst -> [Type] -> [Type]

Substitute within several Types The substitution has to satisfy the invariants described in Note [The substitution invariant].

substScaledTys :: HasDebugCallStack => Subst -> [Scaled Type] -> [Scaled Type]

substTyWith :: HasDebugCallStack => [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst

substTyWithCoVars :: [CoVar] -> [Coercion] -> Type -> Type

Substitute covars within a type

substTyAddInScope :: Subst -> Type -> Type

Substitute within a Type after adding the free variables of the type to the in-scope set. This is useful for the case when the free variables aren't already in the in-scope set or easily available. See also Note [The substitution invariant].

substTyUnchecked :: Subst -> Type -> Type

Substitute within a Type disabling the sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substTysUnchecked :: Subst -> [Type] -> [Type]

Substitute within several Types disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTysUnchecked to substTys and remove this function. Please don't use in new code.

substScaledTyUnchecked :: HasDebugCallStack => Subst -> Scaled Type -> Scaled Type

substThetaUnchecked :: Subst -> ThetaType -> ThetaType

Substitute within a ThetaType disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substThetaUnchecked to substTheta and remove this function. Please don't use in new code.

substTyWithUnchecked :: [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substCoUnchecked :: Subst -> Coercion -> Coercion

Substitute within a Coercion disabling sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substCoWithUnchecked :: [TyVar] -> [Type] -> Coercion -> Coercion

Coercion substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substTheta :: HasDebugCallStack => Subst -> ThetaType -> ThetaType

Substitute within a ThetaType The substitution has to satisfy the invariants described in Note [The substitution invariant].

isUnliftedType :: HasDebugCallStack => Type -> Bool

Is the given type definitely unlifted? See Type for what an unlifted type is.

Panics on representation-polymorphic types; See mightBeUnliftedType for a more approximate predicate that behaves better in the presence of representation polymorphism.

isUnboxedTupleType :: Type -> Bool

isPrimitiveType :: Type -> Bool

Returns true of types that are opaque to Haskell.

coreView :: Type -> Maybe Type

This function strips off the top layer only of a type synonym application (if any) its underlying representation type. Returns Nothing if there is nothing to look through.

This function does not look through type family applications.

By being non-recursive and inlined, this case analysis gets efficiently joined onto the case analysis that the caller is already doing

tyCoVarsOfType :: Type -> TyCoVarSet

tyCoVarsOfTypes :: [Type] -> TyCoVarSet

closeOverKinds :: TyCoVarSet -> TyCoVarSet

tyCoFVsOfType :: Type -> FV

The worker for tyCoFVsOfType and tyCoFVsOfTypeList. The previous implementation used unionVarSet which is O(n+m) and can make the function quadratic. It's exported, so that it can be composed with other functions that compute free variables. See Note [FV naming conventions] in GHC.Utils.FV.

Eta-expanded because that makes it run faster (apparently) See Note [FV eta expansion] in GHC.Utils.FV for explanation.

tyCoFVsOfTypes :: [Type] -> FV

tyCoVarsOfTypeDSet :: Type -> DTyCoVarSet

tyCoFVsOfType that returns free variables of a type in a deterministic set. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in GHC.Utils.FV.

tyCoVarsOfTypesDSet :: [Type] -> DTyCoVarSet

Returns free variables of types, including kind variables as a deterministic set. For type synonyms it does not expand the synonym.

closeOverKindsDSet :: DTyVarSet -> DTyVarSet

Add the kind variables free in the kinds of the tyvars in the given set. Returns a deterministic set.

tyCoVarsOfTypeList :: Type -> [TyCoVar]

tyCoFVsOfType that returns free variables of a type in deterministic order. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in GHC.Utils.FV.

tyCoVarsOfTypesList :: [Type] -> [TyCoVar]

Returns free variables of types, including kind variables as a deterministically ordered list. For type synonyms it does not expand the synonym.

noFreeVarsOfType :: Type -> Bool

pprKind :: Kind -> SDoc

pprParendKind :: Kind -> SDoc

pprSigmaType :: Type -> SDoc

pprType :: Type -> SDoc

pprParendType :: Type -> SDoc

pprTypeApp :: TyCon -> [Type] -> SDoc

pprTheta :: ThetaType -> SDoc

pprParendTheta :: ThetaType -> SDoc

pprThetaArrowTy :: ThetaType -> SDoc

pprClassPred :: Class -> [Type] -> SDoc

pprTCvBndr :: ForAllTyBinder -> SDoc

pprTCvBndrs :: [ForAllTyBinder] -> SDoc

tcTyConVisibilities :: TyCon -> [Bool]

For every arg a tycon can take, the returned list says True if the argument is taken visibly, and False otherwise. Ends with an infinite tail of Trues to allow for oversaturation.

isNextTyConArgVisible :: TyCon -> [Type] -> Bool

If the tycon is applied to the types, is the next argument visible?

isNextArgVisible :: TcType -> Bool

Should this type be applied to a visible argument?