GHC.Tc.Utils.Unify

Description

Type subsumption and unification

Documentation

tcWrapResult :: HsExpr GhcRn -> HsExpr GhcTc -> TcSigmaType -> ExpRhoType -> TcM (HsExpr GhcTc)

tcWrapResultO :: CtOrigin -> HsExpr GhcRn -> HsExpr GhcTc -> TcSigmaType -> ExpRhoType -> TcM (HsExpr GhcTc)

tcWrapResultMono

Arguments

:: HasDebugCallStack

=> HsExpr GhcRn

-> HsExpr GhcTc

-> TcRhoType
Actual; a rho-type, not a sigma-type

-> ExpRhoType
Expected

-> TcM (HsExpr GhcTc)

A version of tcWrapResult to use when the actual type is a rho-type, so nothing to instantiate; just go straight to unify. It means we don't need to pass in a CtOrigin.

tcSubType

Arguments

:: CtOrigin
Used when instantiating

-> UserTypeCtxt
Used when skolemising

-> Maybe TypedThing
The expression that has type actual (if known)

-> TcSigmaType
Actual type

-> ExpRhoType
Expected type

-> TcM HsWrapper

Checks that the actual type is more polymorphic than the expected type.

tcSubTypeSigma :: CtOrigin -> UserTypeCtxt -> TcSigmaType -> TcSigmaType -> TcM HsWrapper

tcSubTypePat :: CtOrigin -> UserTypeCtxt -> ExpSigmaType -> TcSigmaType -> TcM HsWrapper

tcSubTypeDS :: HsExpr GhcRn -> TcRhoType -> TcRhoType -> TcM HsWrapper

A subtype check that performs deep subsumption. See also tcSubTypeMono, for when no instantiation is required.

addSubTypeCtxt :: TcType -> ExpType -> TcM a -> TcM a

tcSubTypeAmbiguity :: UserTypeCtxt -> TcSigmaType -> TcSigmaType -> TcM HsWrapper

tcSubMult :: CtOrigin -> Mult -> Mult -> TcM ()

checkConstraints :: SkolemInfoAnon -> [TcTyVar] -> [EvVar] -> TcM result -> TcM (TcEvBinds, result)

checkTvConstraints :: SkolemInfo -> [TcTyVar] -> TcM result -> TcM result

buildImplicationFor :: TcLevel -> SkolemInfoAnon -> [TcTyVar] -> [EvVar] -> WantedConstraints -> TcM (Bag Implication, TcEvBinds)

buildTvImplication :: SkolemInfoAnon -> [TcTyVar] -> TcLevel -> WantedConstraints -> TcM Implication

emitResidualTvConstraint :: SkolemInfo -> [TcTyVar] -> TcLevel -> WantedConstraints -> TcM ()

data DeepSubsumptionFlag

Constructors

Deep

Shallow

getDeepSubsumptionFlag :: TcM DeepSubsumptionFlag

isRhoTyDS :: DeepSubsumptionFlag -> TcType -> Bool

tcSkolemise :: DeepSubsumptionFlag -> UserTypeCtxt -> TcSigmaType -> (TcRhoType -> TcM result) -> TcM (HsWrapper, result)

tcSkolemiseCompleteSig :: TcCompleteSig -> ([ExpPatType] -> TcRhoType -> TcM result) -> TcM (HsWrapper, result)

The wrapper has type: spec_ty ~> expected_ty See Note [Skolemisation] for the differences between tcSkolemiseCompleteSig and tcTopSkolemise

tcSkolemiseExpectedType :: TcSigmaType -> ([ExpPatType] -> TcRhoType -> TcM result) -> TcM (HsWrapper, result)

dsInstantiate :: CtOrigin -> TcType -> TcM (HsWrapper, Type)

unifyType

Arguments

:: Maybe TypedThing
If present, the thing that has type ty1

-> TcTauType

-> TcTauType

-> TcM TcCoercionN

unifyKind :: Maybe TypedThing -> TcKind -> TcKind -> TcM CoercionN

unifyInvisibleType :: TcTauType -> TcTauType -> TcM TcCoercionN

unifyExprType :: HsExpr GhcRn -> TcType -> TcType -> TcM TcCoercionN

unifyTypeAndEmit :: TypeOrKind -> CtOrigin -> TcType -> TcType -> TcM CoercionN

promoteTcType :: TcLevel -> TcType -> TcM (TcCoercionN, TcType)

swapOverTyVars :: Bool -> TcTyVar -> TcTyVar -> Bool

touchabilityTest :: TcLevel -> TcTyVar -> Bool

This is the key test for untouchability: See Note [Unification preconditions] in GHC.Tc.Utils.Unify and Note [Solve by unification] in GHC.Tc.Solver.Equality

True = the variable is touchable

checkTopShape :: MetaInfo -> TcType -> Bool

checkTopShape checks (TYVAR-TV) Note [Unification preconditions]; returns True if these conditions are satisfied. But see the Note for other preconditions, too.

lhsPriority :: TcTyVar -> Int

data UnifyEnv

Constructors

UE

Fields

u_role :: Role

u_loc :: CtLoc

u_rewriters :: RewriterSet

u_defer :: TcRef (Bag Ct)

u_unified :: Maybe (TcRef [TcTyVar])

updUEnvLoc :: UnifyEnv -> (CtLoc -> CtLoc) -> UnifyEnv

setUEnvRole :: UnifyEnv -> Role -> UnifyEnv

uType :: UnifyEnv -> TcType -> TcType -> TcM CoercionN

mightEqualLater :: InertSet -> TcPredType -> CtLoc -> TcPredType -> CtLoc -> Maybe Subst

makeTypeConcrete :: FastString -> ConcreteTvOrigin -> TcType -> TcM (TcCoercion, Cts)

Try to turn the provided type into a concrete type, by ensuring unfilled metavariables are appropriately marked as concrete.

Returns a coercion whose RHS is a zonked type which is "as concrete as possible", and a collection of Wanted equality constraints that are necessary to make the type concrete.

For example, for an input TYPE a[sk] we will return a coercion with RHS TYPE gamma[conc] together with the Wanted equality constraint a ~# gamma.

INVARIANT: the RHS type of the returned coercion is equal to the input type, up to zonking and the returned Wanted equality constraints.

INVARIANT: if this function returns an empty list of constraints then the RHS type of the returned coercion is concrete, in the sense of Note [Concrete types].

matchExpectedListTy :: TcRhoType -> TcM (TcCoercionN, TcRhoType)

matchExpectedTyConApp :: TyCon -> TcRhoType -> TcM (TcCoercionN, [TcSigmaType])

matchExpectedAppTy :: TcRhoType -> TcM (TcCoercion, (TcSigmaType, TcSigmaType))

matchExpectedFunTys :: ExpectedFunTyOrigin -> UserTypeCtxt -> VisArity -> ExpSigmaType -> ([ExpPatType] -> ExpRhoType -> TcM a) -> TcM (HsWrapper, a)

Use this function to split off arguments types when you have an "expected" type.

This function skolemises at each polytype.

Invariant: this function only applies the provided function to a list of argument types which all have a syntactically fixed RuntimeRep in the sense of Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete. See Note [Return arguments with a fixed RuntimeRep].

matchExpectedFunKind

Arguments

:: TypedThing
type, only for errors

-> Arity
n: number of desired arrows

-> TcKind
fun_kind

-> TcM Coercion
co :: fun_kind ~ (arg1 -> ... -> argn -> res)

Breaks apart a function kind into its pieces.

matchActualFunTy

Arguments

:: ExpectedFunTyOrigin
See Note [Herald for matchExpectedFunTys]

-> Maybe TypedThing
The thing with type TcSigmaType

-> (Arity, TcType)
Total number of value args in the call, and the original function type (Both are used only for error messages)

-> TcRhoType
Type to analyse: a TcRhoType

-> TcM (HsWrapper, Scaled TcSigmaTypeFRR, TcSigmaType)

matchActualFunTy looks for just one function arrow, returning an uninstantiated sigma-type.

Invariant: the returned argument type has a syntactically fixed RuntimeRep in the sense of Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete.

See Note [Return arguments with a fixed RuntimeRep].

matchActualFunTys

Arguments

:: ExpectedFunTyOrigin
See Note [Herald for matchExpectedFunTys]

-> CtOrigin

-> Arity

-> TcSigmaType

-> TcM (HsWrapper, [Scaled TcSigmaTypeFRR], TcRhoType)

Like matchExpectedFunTys, but used when you have an "actual" type, for example in function application.

INVARIANT: the returned argument types all have a syntactically fixed RuntimeRep in the sense of Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete. See Note [Return arguments with a fixed RuntimeRep].

checkTyEqRhs :: Monad m => TyEqFlags m a -> TcType -> m (PuResult a Reduction)

Perform the checks specified by the TyEqFlags on the RHS, in order to enforce the unification preconditions of Note [Unification preconditions].

See Note [checkTyEqRhs].

recurseIntoTyConApp :: Monad m => TyEqFlags m a -> TyCon -> [TcType] -> m (PuResult a Reduction)

recurseIntoFamTyConApp :: Monad m => TyEqFlags m a -> TyCon -> [TcType] -> m (PuResult a Reduction)

data PuResult a b

Constructors

PuFail CheckTyEqResult
Pure unifier failure.

Invariant: the CheckTyEqResult is not cteOK; that it, it specifies a problem.

PuOK (Bag a) b
Pure unifier success.

okCheckRefl :: TcType -> PuResult a Reduction

mapCheck :: Monad m => (x -> m (PuResult a Reduction)) -> [x] -> m (PuResult a Reductions)

data TyEqFlags (m :: Type -> Type) a

Options describing how to deal with a type equality in the eager unifier. See checkTyEqRhs

Constructors

TEFTyFam
TFTyFam: LHS is a type family application Invariant: we are not unifying; see notUnifying_TEFTask

Fields

tefTyFam_occursCheck :: CheckTyEqProblem

The CheckTyEqProblem to report for occurs-check failures (soluble or insoluble)

tefTyFam_tyCon :: TyCon

tefTyFam_args :: [Type]

tef_fam_app :: TyEqFamApp m a

How to deal with type family applications

TEFTyVar
TEFTyVar: LHS is a TyVar.

Fields

tefTyVar_occursCheck :: OccursCheck

Occurs check

tefTyVar_levelCheck :: LevelCheck m

Level check

tefTyVar_concreteCheck :: ConcreteCheck m

Concreteness check

tef_fam_app :: TyEqFamApp m a

How to deal with type family applications

notUnifying_TEFTask :: forall (m :: Type -> Type) a. CheckTyEqProblem -> CanEqLHS -> TyEqFlags m a

Create a "not unifying" TyEqFlags from a CanEqLHS.

See use-case (1) in Note [TyEqFlags].

unifyingLHSMetaTyVar_TEFTask :: CtEvidence -> TcTyVar -> TyEqFlags TcM Ct

Create "unifying" TyEqFlags from a TyVarLHS.

Invariant: the argument TcTyVar is a MetaTv.

defaulting_TEFTask :: TcTyVar -> TyEqFlags TcM a

pureTyEqFlags_LHSMetaTyVar :: TyVar -> TyEqFlags Identity ()

TyEqFlags for a pure call of checkTyEqRhs.

A call to checkTyEqRhs with these TyEqFlags will perform an optimistic check: it will return PuFail if there is definitely a problem, but it might return PuOK if it isn't entirely sure.

mkTEFA_Break :: CtEvidence -> EqRel -> FamAppBreaker a -> TyEqFamApp TcM a

data OccursCheck where

Constructors

OC_None :: OccursCheck
No occurs check.

OC_Check
Do an occurs check between the LHS tyvar and the RHS.

Fields

:: { occurs_tv_name :: Name

The Name to perform an occurs-check on

, occurs_problem :: CheckTyEqProblem

Which CheckTyEqProblem to report for occurs-check failures (soluble or insoluble)

} -> OccursCheck

data LevelCheck (m :: Type -> Type) where

What level check to perform, in a call to the pure unifier? A defunctionalisation of the possible level-check functions Interpreter: tyVarLevelCheck

Constructors

LC_None :: forall (m :: Type -> Type). LevelCheck m
No level check.

LC_Check
Do a level check between the LHS tyvar and the occurrence tyvar.

Fail if the level check fails.

True = lenient check, e.g. the levels have a chance of working out after promotion.

Fields

:: forall (m :: Type -> Type). { lc_lvlc :: TcLevel

, lc_lenient :: Bool

} -> LevelCheck m

LC_Promote
Do a level check between the LHS tyvar and the occurrence tyvar.

If the level check fails, and the occurrence is a unification variable, promote it.

False = don't promote under type families (the common case)

True = promote even under type families (see Note [Defaulting equalities] in GHC.Tc.Solver)

Fields

:: { lc_lvlp :: TcLevel

, lc_deep :: Bool

} -> LevelCheck (IOEnv (Env TcGblEnv TcLclEnv))

data ConcreteCheck (m :: Type -> Type) where

Constructors

CC_None :: forall (m :: Type -> Type). ConcreteCheck m
No concreteness check.

CC_Check :: forall (m :: Type -> Type). ConcreteCheck m
Simple check for concreteness, leniently returning OK if concreteness could be achieved after promotion.

CC_Promote :: ConcreteTvOrigin -> ConcreteCheck (IOEnv (Env TcGblEnv TcLclEnv))
Proper concreteness check: promote non-concrete metavariables, failing if there are any problems.

data TyEqFamApp (m :: Type -> Type) a where

What to do when encountering a type-family application while processing a type equality in the pure unifier.

See Note [Family applications in canonical constraints]

Constructors

TEFA_Recurse :: forall (m :: Type -> Type) a. TyEqFamApp m a
Just recurse

TEFA_Break :: forall a. FamAppBreaker a -> TyEqFamApp (IOEnv (Env TcGblEnv TcLclEnv)) a
Recurse, but replace with cycle breaker if that fails, using the specified FamAppBreaker

data FamAppBreaker a where

A defunctionalisation of family application breaker functions. Interpreter: famAppBreaker

Constructors

BreakGiven :: FamAppBreaker (TcTyVar, TcType)

BreakWanted :: CtEvidence -> TcTyVar -> FamAppBreaker Ct

checkPromoteFreeVars :: CheckTyEqProblem -> Name -> TcLevel -> TyCoVarSet -> TcM CheckTyEqResult

simpleUnifyCheck :: UnifyCheckCaller -> TcLevel -> TcTyVar -> TcType -> SimpleUnifyResult

A fast check for unification. May return "not sure", in which case unification might still be OK, but it'll take more work to do (use the full checkTypeEq).

Rejects if lhs_tv occurs in rhs_ty (occurs check)

Rejects foralls unless lhs_tv is RuntimeUnk (used by GHCi debugger) or is a QL instantiation variable

Rejects a non-concrete type if lhs_tv is concrete

Rejects type families unless fam_ok=True

Does a level-check for type variables, to avoid skolem escape

This function is pretty heavily used, so it's optimised not to allocate

data UnifyCheckCaller

Constructors

UC_OnTheFly

UC_QuickLook

UC_Solver

data SimpleUnifyResult

The result type of simpleUnifyCheck.

Constructors

SUC_CannotUnify
Definitely cannot unify (untouchable variable or incompatible top-shape)

SUC_NotSure
The variable is touchable and the top-shape test passed, but it may or may not be OK to unify

SUC_CanUnify
Definitely OK to unify

fillInferResult :: CtOrigin -> TcType -> InferResult -> TcM HsWrapper

fillInferResultNoInst :: TcType -> InferResult -> TcM TcCoercionN

Fill an InferResult with the given type.

If co = fillInferResult t1 infer_res, then co :: t1 ~# t2, where t2 is the type stored in the ir_ref field of infer_res.

This function enforces the following invariants:

Level invariant. The stored type t2 is at the same level as given by the ir_lvl field.

FRR invariant. Whenever the ir_frr field is IFRR_Check, t2 is guaranteed to have a syntactically fixed RuntimeRep, in the sense of Note [Fixed RuntimeRep] in GHC.Tc.Utils.Concrete.