GHC.Tc.Validity

Documentation

data Rank

data UserTypeCtxt

UserTypeCtxt describes the origin of the polymorphic type in the places where we need an expression to have that type

Constructors

FunSigCtxt Name Bool

InfSigCtxt Name

ExprSigCtxt

KindSigCtxt

StandaloneKindSigCtxt Name

TypeAppCtxt

ConArgCtxt Name

TySynCtxt Name

PatSynCtxt Name

PatSigCtxt

RuleSigCtxt Name

ForSigCtxt Name

DefaultDeclCtxt

InstDeclCtxt Bool

SpecInstCtxt

GenSigCtxt

GhciCtxt Bool

ClassSCCtxt Name

SigmaCtxt

DataTyCtxt Name

DerivClauseCtxt

TyVarBndrKindCtxt Name

DataKindCtxt Name

TySynKindCtxt Name

TyFamResKindCtxt Name

checkValidType :: UserTypeCtxt -> Type -> TcM ()

checkValidMonoType :: Type -> TcM ()

checkValidTheta :: UserTypeCtxt -> ThetaType -> TcM ()

checkValidInstance :: UserTypeCtxt -> LHsSigType GhcRn -> Type -> TcM ()

checkValidInstHead :: UserTypeCtxt -> Class -> [Type] -> TcM ()

validDerivPred :: TyVarSet -> PredType -> Bool

checkTySynRhs :: UserTypeCtxt -> TcType -> TcM ()

checkValidCoAxiom :: CoAxiom Branched -> TcM ()

checkValidCoAxBranch :: TyCon -> CoAxBranch -> TcM ()

checkValidTyFamEqn

Arguments

:: TyCon
of the type family

-> [Var]
Bound variables in the equation

-> [Type]
Type patterns

-> Type
Rhs

-> TcM ()

Do validity checks on a type family equation, including consistency with any enclosing class instance head, termination, and lack of polytypes.

checkValidAssocTyFamDeflt

Arguments

:: TyCon
of the type family

-> [Type]
Type patterns

-> TcM ()

Checks that an associated type family default:

Only consists of arguments that are bare type variables, and

Has a distinct type variable in each argument.

See Note [Type-checking default assoc decls] in GHC.Tc.TyCl.

checkConsistentFamInst

Arguments

:: AssocInstInfo

-> TyCon
Family tycon

-> CoAxBranch

-> TcM ()

badATErr :: Name -> Name -> SDoc

arityErr :: Outputable a => SDoc -> a -> Int -> Int -> SDoc

checkTyConTelescope :: TyCon -> TcM ()

allDistinctTyVars :: TyVarSet -> [KindOrType] -> Bool