GHC.Tc.Zonk.Env

Description

The ZonkEnv zonking environment, and the ZonkT and ZonkBndrT monad transformers, for the final zonking to type in GHC.Tc.Zonk.Type.

See Note [Module structure for zonking] in GHC.Tc.Zonk.Type.

The ZonkEnv

data ZonkEnv

See Note [The ZonkEnv]

Constructors

ZonkEnv

Fields

ze_flexi :: !ZonkFlexi

ze_tv_env :: TyCoVarEnv TyCoVar

ze_id_env :: IdEnv Id

ze_meta_tv_env :: IORef (TyVarEnv Type)

getZonkEnv :: forall (m :: Type -> Type). Monad m => ZonkT m ZonkEnv

data ZonkFlexi

How should we handle unfilled unification variables in the zonker?

See Note [Un-unified unification variables]

Constructors

DefaultFlexi
Default unbound unification variables to Any

SkolemiseFlexi
Skolemise unbound unification variables See Note [Zonking the LHS of a RULE]

RuntimeUnkFlexi
Used in the GHCi debugger

NoFlexi
Panic on unfilled meta-variables See Note [Error on unconstrained meta-variables] in GHC.Tc.Utils.TcMType

initZonkEnv :: MonadIO m => ZonkFlexi -> ZonkT m b -> m b

The ZonkT and ZonkBndrT zonking monad transformers

data ZonkT (m :: Type -> Type) a where

A reader monad over ZonkEnv, for zonking computations which don't modify the ZonkEnv (e.g. don't bind any variables).

Use ZonkBndrT when you need to modify the ZonkEnv (e.g. to bind a variable).

Bundled Patterns

pattern ZonkT :: (ZonkEnv -> m a) -> ZonkT m a

newtype ZonkBndrT (m :: Type -> Type) a

Zonk binders, bringing them into scope in the inner computation.

Can be thought of as a state monad transformer StateT ZonkEnv m a, but written in continuation-passing style.

See Note [Continuation-passing style for zonking].

Constructors

ZonkBndrT

Fields

runZonkBndrT' :: forall r. (a -> ZonkT m r) -> ZonkT m r

Going between ZonkT and ZonkBndrT

runZonkBndrT :: forall (m :: Type -> Type) a. ZonkBndrT m a -> forall r. (a -> ZonkT m r) -> ZonkT m r

Zonk some binders and run the continuation.

Example:

zonk (ForAllTy (Bndr tv vis) body_ty)
 = runZonkBndrT (zonkTyBndrX tv) $ \ tv' ->
   do { body_ty' <- zonkTcTypeToTypeX body_ty
      ; return (ForAllTy (Bndr tv' vis) body_ty') }

See Note [Continuation-passing style for zonking].

noBinders :: forall (m :: Type -> Type) a. Monad m => ZonkT m a -> ZonkBndrT m a

Embed a computation that doesn't modify the ZonkEnv into ZonkBndrT.

don'tBind :: forall (m :: Type -> Type) a. Monad m => ZonkBndrT m a -> ZonkT m a

Run a nested computation that modifies the ZonkEnv, without affecting the outer environment.

Modifying and extending the ZonkEnv in ZonkBndrT

setZonkType :: forall (m :: Type -> Type) a. ZonkFlexi -> ZonkT m a -> ZonkT m a

extendZonkEnv :: forall (m :: Type -> Type). [Var] -> ZonkBndrT m ()

extendIdZonkEnv :: forall (m :: Type -> Type). Var -> ZonkBndrT m ()

extendIdZonkEnvRec :: forall (m :: Type -> Type). [Var] -> ZonkBndrT m ()

Extend the knot-tied environment.

extendTyZonkEnv :: forall (m :: Type -> Type). TyVar -> ZonkBndrT m ()