GHC.Tc.Zonk.Monad

Description

The ZonkM monad, a stripped down TcM, used when zonking within the typechecker in GHC.Tc.Zonk.TcType.

See Note [Module structure for zonking] in GHC.Tc.Zonk.Type.

The ZonkM monad, a stripped down TcM for zonking

data ZonkM a where

A stripped down version of TcM which is sufficient for zonking types.

Bundled Patterns

pattern ZonkM :: (ZonkGblEnv -> IO a) -> ZonkM a

data ZonkGblEnv

Information needed by the ZonkM monad, which is a slimmed down version of TcM with just enough information for zonking.

Constructors

ZonkGblEnv

Fields

zge_logger :: Logger

zge_name_ppr_ctx :: NamePprCtx

zge_src_span :: SrcSpan

zge_tc_level :: TcLevel

zge_binder_stack :: TcBinderStack

getZonkGblEnv :: ZonkM ZonkGblEnv

getZonkTcLevel :: ZonkM TcLevel

Logging within ZonkM

traceZonk :: String -> SDoc -> ZonkM ()

Same as traceTc, but for the ZonkM monad.