GHC.Tc.Zonk.TcType

Description

Zonking types within the typechecker.

Distinct from the final zonking pass in GHC.Tc.Zonk.Type; see Note [Module structure for zonking] in GHC.Tc.Zonk.Type.

Zonking (within the typechecker)

The ZonkM monad and ZonkGblEnv

module GHC.Tc.Zonk.Monad

Zonking types

zonkTcType :: TcType -> ZonkM TcType

zonkTcTypes :: [TcType] -> ZonkM [TcType]

zonkScaledTcType :: Scaled TcType -> ZonkM (Scaled TcType)

zonkTcTyVar :: TcTyVar -> ZonkM TcType

zonkTcTyVars :: [TcTyVar] -> ZonkM [TcType]

zonkTcTyVarToTcTyVar :: HasDebugCallStack => TcTyVar -> ZonkM TcTyVar

zonkTcTyVarsToTcTyVars :: HasDebugCallStack => [TcTyVar] -> ZonkM [TcTyVar]

zonkInvisTVBinder :: VarBndr TcTyVar spec -> ZonkM (VarBndr TcTyVar spec)

zonkCo :: Coercion -> ZonkM Coercion

Zonking TyCons

zonkTcTyCon :: TcTyCon -> ZonkM TcTyCon

FreeVars

zonkTcTypeAndFV :: TcType -> ZonkM DTyCoVarSet

zonkTyCoVarsAndFV :: TyCoVarSet -> ZonkM TyCoVarSet

zonkTyCoVarsAndFVList :: [TyCoVar] -> ZonkM [TyCoVar]

zonkDTyCoVarSetAndFV :: DTyCoVarSet -> ZonkM DTyCoVarSet

Zonking CoVars and Ids

zonkId :: TcId -> ZonkM TcId

zonkId is used *during* typechecking just to zonk the Id's type

zonkCoVar :: CoVar -> ZonkM CoVar

zonkTyCoVar :: TyCoVar -> ZonkM TcType

zonkTyCoVarKind :: TyCoVar -> ZonkM TyCoVar

zonkTyCoVarBndrKind :: VarBndr TyCoVar flag -> ZonkM (VarBndr TyCoVar flag)

Zonking skolem info

zonkSkolemInfo :: SkolemInfo -> ZonkM SkolemInfo

zonkSkolemInfoAnon :: SkolemInfoAnon -> ZonkM SkolemInfoAnon

Zonking constraints

zonkCt :: Ct -> ZonkM Ct

zonkWC :: WantedConstraints -> ZonkM WantedConstraints

zonkSimples :: Cts -> ZonkM Cts

zonkImplication :: Implication -> ZonkM Implication

Tidying

tcInitTidyEnv :: ZonkM TidyEnv

tcInitOpenTidyEnv :: [TyCoVar] -> ZonkM TidyEnv

Get a TidyEnv that includes mappings for all vars free in the given type. Useful when tidying open types.

tidyCt :: TidyEnv -> Ct -> Ct

tidyEvVar :: TidyEnv -> EvVar -> EvVar

tidyDelayedError :: TidyEnv -> DelayedError -> DelayedError

Zonk & tidy

zonkTidyTcType :: TidyEnv -> TcType -> ZonkM (TidyEnv, TcType)

zonkTidyTcTypes :: TidyEnv -> [TcType] -> ZonkM (TidyEnv, [TcType])

zonkTidyOrigin :: TidyEnv -> CtOrigin -> ZonkM (TidyEnv, CtOrigin)

zonkTidyOrigins :: TidyEnv -> [CtOrigin] -> ZonkM (TidyEnv, [CtOrigin])

zonkTidyFRRInfos :: TidyEnv -> [FixedRuntimeRepErrorInfo] -> ZonkM (TidyEnv, [FixedRuntimeRepErrorInfo])

Writing to metavariables

writeMetaTyVar

Arguments

:: HasDebugCallStack

=> TcTyVar
the type varfiable to write to

-> TcType
the type to write into the mutable reference

-> ZonkM ()

Write into a currently-empty MetaTyVar.

Works with both type and kind variables.

writeMetaTyVarRef

Arguments

:: HasDebugCallStack

=> TcTyVar
for debug assertions only;

-> TcRef MetaDetails
ref cell must be for the same tyvar

-> TcType
the type to write to the mutable reference

-> ZonkM ()

Write into the MetaDetails mutable references of a MetaTv.

Handling coercion holes

checkCoercionHole :: CoVar -> Coercion -> ZonkM Coercion

Check that a coercion is appropriate for filling a hole. (The hole itself is needed only for printing.) Always returns the checked coercion, but this return value is necessary so that the input coercion is forced only when the output is forced.