GHC.Types.Avail

Documentation

type Avails = [AvailInfo]

A collection of AvailInfo - several things that are "available"

data AvailInfo

Records what things are "available", i.e. in scope

Constructors

Avail GreName
An ordinary identifier in scope, or a field label without a parent type (see Note [Representing pattern synonym fields in AvailInfo]).

AvailTC
A type or class in scope

The AvailTC Invariant: If the type or class is itself to be in scope, it must be first in this list. Thus, typically:

AvailTC Eq [Eq, ==, \/=]

Fields

Name

The name of the type or class

[GreName]

The available pieces of type or class (see Note [Representing fields in AvailInfo]).

avail :: Name -> AvailInfo

availField :: FieldLabel -> AvailInfo

availTC :: Name -> [Name] -> [FieldLabel] -> AvailInfo

availsToNameSet :: [AvailInfo] -> NameSet

availsToNameSetWithSelectors :: [AvailInfo] -> NameSet

availsToNameEnv :: [AvailInfo] -> NameEnv AvailInfo

availExportsDecl :: AvailInfo -> Bool

Does this AvailInfo export the parent decl? This depends on the invariant that the parent is first if it appears at all.

availName :: AvailInfo -> Name

Just the main name made available, i.e. not the available pieces of type or class brought into scope by the AvailInfo

availGreName :: AvailInfo -> GreName

availNames :: AvailInfo -> [Name]

All names made available by the availability information (excluding overloaded selectors)

availNonFldNames :: AvailInfo -> [Name]

Names for non-fields made available by the availability information

availNamesWithSelectors :: AvailInfo -> [Name]

All names made available by the availability information (including overloaded selectors)

availFlds :: AvailInfo -> [FieldLabel]

Fields made available by the availability information

availGreNames :: AvailInfo -> [GreName]

Names and fields made available by the availability information.

availSubordinateGreNames :: AvailInfo -> [GreName]

Names and fields made available by the availability information, other than the main decl itself.

stableAvailCmp :: AvailInfo -> AvailInfo -> Ordering

Compare lexicographically

plusAvail :: AvailInfo -> AvailInfo -> AvailInfo

trimAvail :: AvailInfo -> Name -> AvailInfo

trims an AvailInfo to keep only a single name

filterAvail :: (Name -> Bool) -> AvailInfo -> [AvailInfo] -> [AvailInfo]

filters an AvailInfo by the given predicate

filterAvails :: (Name -> Bool) -> [AvailInfo] -> [AvailInfo]

filters AvailInfos by the given predicate

nubAvails :: [AvailInfo] -> [AvailInfo]

Combines AvailInfos from the same family avails may have several items with the same availName E.g import Ix( Ix(..), index ) will give Ix(Ix,index,range) and Ix(index) We want to combine these; addAvail does that

data GreName

Used where we may have an ordinary name or a record field label. See Note [GreNames] in GHC.Types.Name.Reader.

Constructors

NormalGreName Name

FieldGreName FieldLabel

greNameMangledName :: GreName -> Name

A Name for internal use, but not for output to the user. For fields, the OccName will be the selector. See Note [GreNames] in GHC.Types.Name.Reader.

greNamePrintableName :: GreName -> Name

A Name suitable for output to the user. For fields, the OccName will be the field label. See Note [GreNames] in GHC.Types.Name.Reader.

greNameSrcSpan :: GreName -> SrcSpan

greNameFieldLabel :: GreName -> Maybe FieldLabel

partitionGreNames :: [GreName] -> ([Name], [FieldLabel])

stableGreNameCmp :: GreName -> GreName -> Ordering