GHC.Types.Basic

Documentation

data LeftOrRight

Constructors

CLeft

CRight

pickLR :: LeftOrRight -> (a, a) -> a

type ConTag = Int

Constructor Tag

Type of the tags associated with each constructor possibility or superclass selector

type ConTagZ = Int

A *zero-indexed* constructor tag

fIRST_TAG :: ConTag

Tags are allocated from here for real constructors or for superclass selectors

type Arity = Int

The number of value arguments that can be applied to a value before it does "real work". So: fib 100 has arity 0 x -> fib x has arity 1 See also Note [Definition of arity] in GHC.Core.Opt.Arity

type RepArity = Int

Representation Arity

The number of represented arguments that can be applied to a value before it does "real work". So: fib 100 has representation arity 0 x -> fib x has representation arity 1 (# x, y #) -> fib (x + y) has representation arity 2

type JoinArity = Int

The number of arguments that a join point takes. Unlike the arity of a function, this is a purely syntactic property and is fixed when the join point is created (or converted from a value). Both type and value arguments are counted.

data Alignment

A power-of-two alignment

mkAlignment :: Int -> Alignment

alignmentOf :: Int -> Alignment

alignmentBytes :: Alignment -> Int

data PromotionFlag

Is a TyCon a promoted data constructor or just a normal type constructor?

Constructors

NotPromoted

IsPromoted

isPromoted :: PromotionFlag -> Bool

data FunctionOrData

Constructors

IsFunction

IsData

data WarningTxt

Warning Text

reason/explanation from a WARNING or DEPRECATED pragma

Constructors

WarningTxt (Located SourceText) [Located StringLiteral]

DeprecatedTxt (Located SourceText) [Located StringLiteral]

pprWarningTxtForMsg :: WarningTxt -> SDoc

data StringLiteral

A String Literal in the source, including its original raw format for use by source to source manipulation tools.

Constructors

StringLiteral

Fields

sl_st :: SourceText

sl_fs :: FastString

data Fixity

Constructors

Fixity SourceText Int FixityDirection

data FixityDirection

Constructors

InfixL

InfixR

InfixN

defaultFixity :: Fixity

maxPrecedence :: Int

minPrecedence :: Int

negateFixity :: Fixity

funTyFixity :: Fixity

compareFixity :: Fixity -> Fixity -> (Bool, Bool)

data LexicalFixity

Captures the fixity of declarations as they are parsed. This is not necessarily the same as the fixity declaration, as the normal fixity may be overridden using parens or backticks.

Constructors

Prefix

Infix

data RecFlag

Recursivity Flag

Constructors

Recursive

NonRecursive

isRec :: RecFlag -> Bool

isNonRec :: RecFlag -> Bool

boolToRecFlag :: Bool -> RecFlag

data Origin

Constructors

FromSource

Generated

isGenerated :: Origin -> Bool

type RuleName = FastString

pprRuleName :: RuleName -> SDoc

data TopLevelFlag

Constructors

TopLevel

NotTopLevel

isTopLevel :: TopLevelFlag -> Bool

isNotTopLevel :: TopLevelFlag -> Bool

data OverlapFlag

The semantics allowed for overlapping instances for a particular instance. See Note [Safe Haskell isSafeOverlap] (in GHC.Core.InstEnv) for a explanation of the isSafeOverlap field.

AnnKeywordId : AnnOpen '{-# OVERLAPPABLE' or '{-# OVERLAPPING' or '{-# OVERLAPS' or '{-# INCOHERENT', AnnClose `#-}`,

Constructors

OverlapFlag

Fields

overlapMode :: OverlapMode

isSafeOverlap :: Bool

data OverlapMode

Constructors

NoOverlap SourceText
This instance must not overlap another NoOverlap instance. However, it may be overlapped by Overlapping instances, and it may overlap Overlappable instances.

Overlappable SourceText
Silently ignore this instance if you find a more specific one that matches the constraint you are trying to resolve

Example: constraint (Foo [Int]) instance Foo [Int] instance {-# OVERLAPPABLE #-} Foo [a]

Since the second instance has the Overlappable flag, the first instance will be chosen (otherwise its ambiguous which to choose)

Overlapping SourceText
Silently ignore any more general instances that may be used to solve the constraint.

Example: constraint (Foo [Int]) instance {-# OVERLAPPING #-} Foo [Int] instance Foo [a]

Since the first instance has the Overlapping flag, the second---more general---instance will be ignored (otherwise it is ambiguous which to choose)

Overlaps SourceText
Equivalent to having both Overlapping and Overlappable flags.

Incoherent SourceText
Behave like Overlappable and Overlapping, and in addition pick an arbitrary one if there are multiple matching candidates, and don't worry about later instantiation

Example: constraint (Foo [b]) instance {-# INCOHERENT -} Foo [Int] instance Foo [a] Without the Incoherent flag, we'd complain that instantiating b would change which instance was chosen. See also note [Incoherent instances] in GHC.Core.InstEnv

setOverlapModeMaybe :: OverlapFlag -> Maybe OverlapMode -> OverlapFlag

hasOverlappingFlag :: OverlapMode -> Bool

hasOverlappableFlag :: OverlapMode -> Bool

hasIncoherentFlag :: OverlapMode -> Bool

data Boxity

Constructors

Boxed

Unboxed

isBoxed :: Boxity -> Bool

newtype PprPrec

A general-purpose pretty-printing precedence type.

Constructors

PprPrec Int

topPrec :: PprPrec

sigPrec :: PprPrec

opPrec :: PprPrec

funPrec :: PprPrec

starPrec :: PprPrec

appPrec :: PprPrec

maybeParen :: PprPrec -> PprPrec -> SDoc -> SDoc

data TupleSort

Constructors

BoxedTuple

UnboxedTuple

ConstraintTuple

tupleSortBoxity :: TupleSort -> Boxity

boxityTupleSort :: Boxity -> TupleSort

tupleParens :: TupleSort -> SDoc -> SDoc

sumParens :: SDoc -> SDoc

pprAlternative

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> a
The things to be pretty printed

-> ConTag
Alternative (one-based)

-> Arity
Arity

-> SDoc
SDoc where the alternative havs been pretty printed and finally packed into a paragraph.

Pretty print an alternative in an unboxed sum e.g. "| a | |".

The OneShotInfo type

data OneShotInfo

If the Id is a lambda-bound variable then it may have lambda-bound variable info. Sometimes we know whether the lambda binding this variable is a "one-shot" lambda; that is, whether it is applied at most once.

This information may be useful in optimisation, as computations may safely be floated inside such a lambda without risk of duplicating work.

See also Note [OneShotInfo overview] above.

Constructors

NoOneShotInfo
No information

OneShotLam
The lambda is applied at most once.

noOneShotInfo :: OneShotInfo

It is always safe to assume that an Id has no lambda-bound variable information

hasNoOneShotInfo :: OneShotInfo -> Bool

isOneShotInfo :: OneShotInfo -> Bool

bestOneShot :: OneShotInfo -> OneShotInfo -> OneShotInfo

worstOneShot :: OneShotInfo -> OneShotInfo -> OneShotInfo

data OccInfo

identifier Occurrence Information

Constructors

ManyOccs
There are many occurrences, or unknown occurrences

Fields

occ_tail :: !TailCallInfo

IAmDead
Marks unused variables. Sometimes useful for lambda and case-bound variables.

OneOcc
Occurs exactly once (per branch), not inside a rule

Fields

occ_in_lam :: !InsideLam

occ_n_br :: !BranchCount

occ_int_cxt :: !InterestingCxt

occ_tail :: !TailCallInfo

IAmALoopBreaker
This identifier breaks a loop of mutually recursive functions. The field marks whether it is only a loop breaker due to a reference in a rule

Fields

occ_rules_only :: !RulesOnly

occ_tail :: !TailCallInfo

noOccInfo :: OccInfo

seqOccInfo :: OccInfo -> ()

zapFragileOcc :: OccInfo -> OccInfo

isOneOcc :: OccInfo -> Bool

isDeadOcc :: OccInfo -> Bool

isStrongLoopBreaker :: OccInfo -> Bool

isWeakLoopBreaker :: OccInfo -> Bool

isManyOccs :: OccInfo -> Bool

isNoOccInfo :: OccInfo -> Bool

strongLoopBreaker :: OccInfo

weakLoopBreaker :: OccInfo

data InsideLam

Inside Lambda

Constructors

IsInsideLam
Occurs inside a non-linear lambda Substituting a redex for this occurrence is dangerous because it might duplicate work.

NotInsideLam

type BranchCount = Int

oneBranch :: BranchCount

data InterestingCxt

Interesting Context

Constructors

IsInteresting
Function: is applied Data value: scrutinised by a case with at least one non-DEFAULT branch

NotInteresting

data TailCallInfo

Constructors

AlwaysTailCalled JoinArity

NoTailCallInfo

tailCallInfo :: OccInfo -> TailCallInfo

zapOccTailCallInfo :: OccInfo -> OccInfo

isAlwaysTailCalled :: OccInfo -> Bool

data EP a

Embedding Projection pair

Constructors

EP

Fields

fromEP :: a

toEP :: a

data DefMethSpec ty

Default Method Specification

Constructors

VanillaDM

GenericDM ty

data SwapFlag

Constructors

NotSwapped

IsSwapped

flipSwap :: SwapFlag -> SwapFlag

unSwap :: SwapFlag -> (a -> a -> b) -> a -> a -> b

isSwapped :: SwapFlag -> Bool

data CompilerPhase

Constructors

InitialPhase

Phase PhaseNum

FinalPhase

type PhaseNum = Int

Phase Number

data Activation

Constructors

AlwaysActive

ActiveBefore SourceText PhaseNum

ActiveAfter SourceText PhaseNum

FinalActive

NeverActive

isActive :: CompilerPhase -> Activation -> Bool

competesWith :: Activation -> Activation -> Bool

isNeverActive :: Activation -> Bool

isAlwaysActive :: Activation -> Bool

activeInFinalPhase :: Activation -> Bool

activateAfterInitial :: Activation

activateDuringFinal :: Activation

data RuleMatchInfo

Rule Match Information

Constructors

ConLike

FunLike

isConLike :: RuleMatchInfo -> Bool

isFunLike :: RuleMatchInfo -> Bool

data InlineSpec

Inline Specification

Constructors

Inline

Inlinable

NoInline

NoUserInline

noUserInlineSpec :: InlineSpec -> Bool

data InlinePragma

Constructors

InlinePragma

Fields

inl_src :: SourceText

inl_inline :: InlineSpec

inl_sat :: Maybe Arity

inl_act :: Activation

inl_rule :: RuleMatchInfo

defaultInlinePragma :: InlinePragma

alwaysInlinePragma :: InlinePragma

neverInlinePragma :: InlinePragma

dfunInlinePragma :: InlinePragma

isDefaultInlinePragma :: InlinePragma -> Bool

isInlinePragma :: InlinePragma -> Bool

isInlinablePragma :: InlinePragma -> Bool

isAnyInlinePragma :: InlinePragma -> Bool

inlinePragmaSpec :: InlinePragma -> InlineSpec

inlinePragmaSat :: InlinePragma -> Maybe Arity

inlinePragmaActivation :: InlinePragma -> Activation

inlinePragmaRuleMatchInfo :: InlinePragma -> RuleMatchInfo

setInlinePragmaActivation :: InlinePragma -> Activation -> InlinePragma

setInlinePragmaRuleMatchInfo :: InlinePragma -> RuleMatchInfo -> InlinePragma

pprInline :: InlinePragma -> SDoc

pprInlineDebug :: InlinePragma -> SDoc

data SuccessFlag

Constructors

Succeeded

Failed

succeeded :: SuccessFlag -> Bool

failed :: SuccessFlag -> Bool

successIf :: Bool -> SuccessFlag

data IntegralLit

Integral Literal

Used (instead of Integer) to represent negative zegative zero which is required for NegativeLiterals extension to correctly parse `-0::Double` as negative zero. See also #13211.

Constructors

IL

Fields

il_text :: SourceText

il_neg :: Bool

il_value :: Integer

data FractionalLit

Fractional Literal

Used (instead of Rational) to represent exactly the floating point literal that we encountered in the user's source program. This allows us to pretty-print exactly what the user wrote, which is important e.g. for floating point numbers that can't represented as Doubles (we used to via Double for pretty-printing). See also #2245.

Constructors

FL

Fields

fl_text :: SourceText

fl_neg :: Bool

fl_value :: Rational

negateIntegralLit :: IntegralLit -> IntegralLit

negateFractionalLit :: FractionalLit -> FractionalLit

mkIntegralLit :: Integral a => a -> IntegralLit

mkFractionalLit :: Real a => a -> FractionalLit

integralFractionalLit :: Bool -> Integer -> FractionalLit

data SourceText

Constructors

SourceText String

NoSourceText
For when code is generated, e.g. TH, deriving. The pretty printer will then make its own representation of the item.

pprWithSourceText :: SourceText -> SDoc -> SDoc

Special combinator for showing string literals.

data IntWithInf

An integer or infinity

infinity :: IntWithInf

A representation of infinity

treatZeroAsInf :: Int -> IntWithInf

Turn a positive number into an IntWithInf, where 0 represents infinity

mkIntWithInf :: Int -> IntWithInf

Inject any integer into an IntWithInf

intGtLimit :: Int -> IntWithInf -> Bool

data SpliceExplicitFlag

Constructors

ExplicitSplice
= $(f x y)

ImplicitSplice
= f x y, i.e. a naked top level expression

data TypeOrKind

Flag to see whether we're type-checking terms or kind-checking types

Constructors

TypeLevel

KindLevel

isTypeLevel :: TypeOrKind -> Bool

isKindLevel :: TypeOrKind -> Bool