GHC.Types.Cpr

Description

Types for the Constructed Product Result lattice. GHC.Core.Opt.CprAnal and GHC.Core.Opt.WorkWrap.Utils are its primary customers via idCprInfo.

Documentation

data Cpr where

Bundled Patterns

pattern ConCpr :: ConTag -> [Cpr] -> Cpr

topCpr :: Cpr

botCpr :: Cpr

flatConCpr :: ConTag -> Cpr

asConCpr :: Cpr -> Maybe (ConTag, [Cpr])

data CprType

The abstract domain \(A_t\) from the original 'CPR for Haskell' paper.

Constructors

CprType

Fields

ct_arty :: !Arity

Number of value arguments the denoted expression eats before returning the ct_cpr

ct_cpr :: !Cpr

Cpr eventually unleashed when applied to ct_arty arguments

topCprType :: CprType

botCprType :: CprType

flatConCprType :: ConTag -> CprType

lubCprType :: CprType -> CprType -> CprType

applyCprTy :: CprType -> Arity -> CprType

abstractCprTy :: CprType -> CprType

trimCprTy :: CprType -> CprType

data UnpackConFieldsResult

The result of unpackConFieldsCpr.

Constructors

AllFieldsSame !Cpr

ForeachField ![Cpr]

unpackConFieldsCpr :: DataCon -> Cpr -> UnpackConFieldsResult

Unpacks a ConCpr-shaped Cpr and returns the field Cprs wrapped in a ForeachField. Otherwise, it returns AllFieldsSame with the appropriate Cpr to assume for each field.

The use of UnpackConFieldsResult allows O(1) space for the common, non-ConCpr case.

newtype CprSig

The arity of the wrapped CprType is the arity at which it is safe to unleash. See Note [Understanding DmdType and StrictSig] in GHC.Types.Demand

Constructors

CprSig

Fields

getCprSig :: CprType

topCprSig :: CprSig

isTopCprSig :: CprSig -> Bool

mkCprSigForArity :: Arity -> CprType -> CprSig

Turns a CprType computed for the particular Arity into a CprSig unleashable at that arity. See Note [Understanding DmdType and StrictSig] in GHC.Types.Demand

mkCprSig :: Arity -> Cpr -> CprSig

seqCprSig :: CprSig -> ()