GHC.Types.DefaultEnv

Documentation

data ClassDefaults

Defaulting type assignments for the given class.

Constructors

ClassDefaults

Fields

cd_class :: Class

The class whose defaults are being defined

cd_types :: [Type]

cd_provenance :: DefaultProvenance

Where the defaults came from see Note [Default exports] in GHC.Tc.Gen.Export

cd_warn :: Maybe (WarningTxt GhcRn)

Warning emitted when the default is used

data DefaultProvenance

The provenance of a collection of default types for a class. see Note [DefaultProvenance] for more details

Constructors

DP_Local
A locally defined default declaration.

Fields

defaultDeclLoc :: SrcSpan

The SrcSpan of the default declaration

defaultDeclH98 :: Bool

Is this a Haskell 98 default declaration?

DP_Builtin
Built-in class defaults.

DP_Imported
Imported class defaults.

Fields

Module

The module from which the defaults were imported

type DefaultEnv = NameEnv ClassDefaults

Default environment mapping class name Name to their default type lists

NB: this includes Haskell98 default declarations, at the Num key.

emptyDefaultEnv :: DefaultEnv

isEmptyDefaultEnv :: DefaultEnv -> Bool

defaultEnv :: [ClassDefaults] -> DefaultEnv

unitDefaultEnv :: ClassDefaults -> DefaultEnv

lookupDefaultEnv :: DefaultEnv -> Name -> Maybe ClassDefaults

filterDefaultEnv :: (ClassDefaults -> Bool) -> DefaultEnv -> DefaultEnv

defaultList :: DefaultEnv -> [ClassDefaults]

plusDefaultEnv :: DefaultEnv -> DefaultEnv -> DefaultEnv

mkDefaultEnv :: [(Name, ClassDefaults)] -> DefaultEnv

Create a DefaultEnv from a list of (Name, ClassDefaults) pairs it is useful if we don't want to poke into the ClassDefaults structure to get the Name of the class, it can be problematic, see #25858

insertDefaultEnv :: ClassDefaults -> DefaultEnv -> DefaultEnv

isHaskell2010Default :: DefaultProvenance -> Bool