GHC.Types.FieldLabel

Documentation

type FieldLabelString = FastString

Field labels are just represented as strings; they are not necessarily unique (even within a module)

type FieldLabelEnv = DFastStringEnv FieldLabel

A map from labels to all the auxiliary information

data FieldLbl a

Fields in an algebraic record type

Constructors

FieldLabel

Fields

flLabel :: FieldLabelString

User-visible label of the field

flIsOverloaded :: Bool

Was DuplicateRecordFields on in the defining module for this datatype?

flSelector :: a

Record selector function

type FieldLabel = FieldLbl Name

mkFieldLabelOccs :: FieldLabelString -> OccName -> Bool -> FieldLbl OccName

Record selector OccNames are built from the underlying field name and the name of the first data constructor of the type, to support duplicate record field names. See Note [Why selector names include data constructors].