GHC.Types.Fixity

Description

Fixity

Documentation

data Fixity

Constructors

Fixity Int FixityDirection

data FixityDirection

Constructors

InfixL

InfixR

InfixN

data LexicalFixity

Captures the fixity of declarations as they are parsed. This is not necessarily the same as the fixity declaration, as the normal fixity may be overridden using parens or backticks.

Constructors

Prefix

Infix

maxPrecedence :: Int

minPrecedence :: Int

defaultFixity :: Fixity

negateFixity :: Fixity

funTyFixity :: Fixity

compareFixity :: Fixity -> Fixity -> (Bool, Bool)