GHC.Types.ForeignCall

Documentation

newtype ForeignCall

Constructors

CCall CCallSpec

isSafeForeignCall :: ForeignCall -> Bool

data Safety

Constructors

PlaySafe
Might invoke Haskell GC, or do a call back, or switch threads, etc. So make sure things are tidy before the call. Additionally, in the threaded RTS we arrange for the external call to be executed by a separate OS thread, i.e., _concurrently_ to the execution of other Haskell threads.

PlayInterruptible
Like PlaySafe, but additionally the worker thread running this foreign call may be unceremoniously killed, so it must be scheduled on an unbound thread.

PlayRisky
None of the above can happen; the call will return without interacting with the runtime system at all. Specifically:

No GC

No call backs

No blocking

No precise exceptions

playSafe :: Safety -> Bool

playInterruptible :: Safety -> Bool

data CExportSpec

Constructors

CExportStatic SourceText CLabelString CCallConv

type CLabelString = FastString

isCLabelString :: CLabelString -> Bool

pprCLabelString :: CLabelString -> SDoc

data CCallSpec

Constructors

CCallSpec CCallTarget CCallConv Safety

data CCallTarget

How to call a particular function in C-land.

Constructors

StaticTarget SourceText CLabelString (Maybe Unit) Bool

DynamicTarget

isDynamicTarget :: CCallTarget -> Bool

data CCallConv

Constructors

CCallConv

CApiConv

StdCallConv

PrimCallConv

JavaScriptCallConv

defaultCCallConv :: CCallConv

ccallConvToInt :: CCallConv -> Int

ccallConvAttribute :: CCallConv -> SDoc

data Header

Constructors

Header SourceText FastString

data CType

A C type, used in CAPI FFI calls

AnnKeywordId : AnnOpen '{-# CTYPE', AnnHeader,AnnVal, AnnClose '#-}',

Constructors

CType SourceText (Maybe Header) (SourceText, FastString)