GHC.Types.HpcInfo

Description

Haskell Program Coverage (HPC) support

Documentation

data HpcInfo

Information about a modules use of Haskell Program Coverage

Constructors

HpcInfo

Fields

hpcInfoTickCount :: Int

hpcInfoHash :: Int

NoHpcInfo

Fields

hpcUsed :: AnyHpcUsage

Is hpc used anywhere on the module *tree*?

type AnyHpcUsage = Bool

This is used to signal if one of my imports used HPC instrumentation even if there is no module-local HPC usage

emptyHpcInfo :: AnyHpcUsage -> HpcInfo

isHpcUsed :: HpcInfo -> AnyHpcUsage

Find out if HPC is used by this module or any of the modules it depends upon