GHC.Types.IPE

Documentation

type DCMap = UniqMap DataCon (NonEmpty (Int, Maybe (RealSrcSpan, String)))

A map storing all the different uses of a specific data constructor and the approximate source position that usage arose from. The Int is an incrementing identifier which distinguishes each usage of a constructor in a module. It is paired with the source position the constructor was used at, if possible and a string which names the source location. This is the same information as is the payload for the SourceNote constructor.

type ClosureMap = UniqMap Name (Type, Maybe (RealSrcSpan, String))

A map from a Name to the best approximate source position that name arose from.

data InfoTableProvMap

Constructors

InfoTableProvMap

Fields

provDC :: DCMap

provClosure :: ClosureMap

emptyInfoTableProvMap :: InfoTableProvMap