GHC.Types.Id

Description

GHC uses several kinds of name internally:

OccName: see GHC.Types.Name.Occurrence

RdrName: see GHC.Types.Name.Reader

Name: see GHC.Types.Name

Id represents names that not only have a Name but also a Type and some additional details (a IdInfo and one of LocalIdDetails or GlobalIdDetails) that are added, modified and inspected by various compiler passes. These Var names may either be global or local, see GHC.Types.Var

Var: see GHC.Types.Var

The main types

data Var

Variable

Essentially a typed Name, that may also contain some additional information about the Var and its use sites.

type Id = Var

Identifier

isId :: Var -> Bool

Is this a value-level (i.e., computationally relevant) Varentifier? Satisfies isId = not . isTyVar.

In and Out variants

type InVar = Var

type InId = Id

type OutVar = Var

type OutId = Id

Simple construction

mkGlobalId :: IdDetails -> Name -> Type -> IdInfo -> Id

For an explanation of global vs. local Ids, see GHC.Types.Var.Var

mkVanillaGlobal :: Name -> Type -> Id

Make a global Id without any extra information at all

mkVanillaGlobalWithInfo :: Name -> Type -> IdInfo -> Id

Make a global Id with no global information but some generic IdInfo

mkLocalId :: HasDebugCallStack => Name -> Mult -> Type -> Id

For an explanation of global vs. local Ids, see GHC.Types.Var

mkLocalCoVar :: Name -> Type -> CoVar

Make a local CoVar

mkLocalIdOrCoVar :: Name -> Mult -> Type -> Id

Like mkLocalId, but checks the type to see if it should make a covar

mkLocalIdWithInfo :: HasDebugCallStack => Name -> Mult -> Type -> IdInfo -> Id

mkExportedLocalId :: IdDetails -> Name -> Type -> Id

Create a local Id that is marked as exported. This prevents things attached to it from being removed as dead code. See Note [Exported LocalIds]

mkExportedVanillaId :: Name -> Type -> Id

mkSysLocal :: FastString -> Unique -> Mult -> Type -> Id

Create a system local Id. These are local Ids (see Var) that are created by the compiler out of thin air

mkSysLocalM :: MonadUnique m => FastString -> Mult -> Type -> m Id

mkSysLocalOrCoVar :: FastString -> Unique -> Mult -> Type -> Id

Like mkSysLocal, but checks to see if we have a covar type

mkSysLocalOrCoVarM :: MonadUnique m => FastString -> Mult -> Type -> m Id

mkUserLocal :: OccName -> Unique -> Mult -> Type -> SrcSpan -> Id

Create a user local Id. These are local Ids (see GHC.Types.Var) with a name and location that the user might recognize

mkUserLocalOrCoVar :: OccName -> Unique -> Mult -> Type -> SrcSpan -> Id

Like mkUserLocal, but checks if we have a coercion type

mkTemplateLocals :: [Type] -> [Id]

Create a template local for a series of types

mkTemplateLocalsNum :: Int -> [Type] -> [Id]

Create a template local for a series of type, but start from a specified template local

mkTemplateLocal :: Int -> Type -> Id

Create a template local: a family of system local Ids in bijection with Ints, typically used in unfoldings

mkScaledTemplateLocal :: Int -> Scaled Type -> Id

mkWorkerId :: Unique -> Id -> Type -> Id

Workers get local names. CoreTidy will externalise these if necessary

Taking an Id apart

idName :: Id -> Name

idType :: Id -> Kind

idMult :: Id -> Mult

idScaledType :: Id -> Scaled Type

idUnique :: Id -> Unique

idInfo :: HasDebugCallStack => Id -> IdInfo

idDetails :: Id -> IdDetails

recordSelectorTyCon :: Id -> RecSelParent

If the Id is that for a record selector, extract the sel_tycon. Panic otherwise.

recordSelectorTyCon_maybe :: Id -> Maybe RecSelParent

Modifying an Id

setIdName :: Id -> Name -> Id

setIdUnique :: Id -> Unique -> Id

setIdType :: Id -> Type -> Id

Not only does this set the Id Type, it also evaluates the type to try and reduce space usage

setIdMult :: Id -> Mult -> Id

updateIdTypeButNotMult :: (Type -> Type) -> Id -> Id

updateIdTypeAndMult :: (Type -> Type) -> Id -> Id

updateIdTypeAndMultM :: Monad m => (Type -> m Type) -> Id -> m Id

setIdExported :: Id -> Id

setIdNotExported :: Id -> Id

globaliseId :: Id -> Id

If it's a local, make it global

localiseId :: Id -> Id

setIdInfo :: Id -> IdInfo -> Id

lazySetIdInfo :: Id -> IdInfo -> Id

modifyIdInfo :: HasDebugCallStack => (IdInfo -> IdInfo) -> Id -> Id

maybeModifyIdInfo :: Maybe IdInfo -> Id -> Id

zapLamIdInfo :: Id -> Id

zapIdDemandInfo :: Id -> Id

zapIdUsageInfo :: Id -> Id

zapIdUsageEnvInfo :: Id -> Id

zapIdUsedOnceInfo :: Id -> Id

zapIdTailCallInfo :: Id -> Id

zapFragileIdInfo :: Id -> Id

zapIdStrictness :: Id -> Id

zapStableUnfolding :: Id -> Id

transferPolyIdInfo :: Id -> [Var] -> Id -> Id

scaleIdBy :: Mult -> Id -> Id

scaleVarBy :: Mult -> Var -> Var

Like scaleIdBy, but skips non-Ids. Useful for scaling a mixed list of ids and tyvars.

Predicates on Ids

isImplicitId :: Id -> Bool

isImplicitId tells whether an Ids info is implied by other declarations, so we don't need to put its signature in an interface file, even if it's mentioned in some other interface unfolding.

isDeadBinder :: Id -> Bool

isStrictId :: Id -> Bool

This predicate says whether the Id has a strict demand placed on it or has a type such that it can always be evaluated strictly (i.e an unlifted type, as of GHC 7.6). We need to check separately whether the Id has a so-called "strict type" because if the demand for the given id hasn't been computed yet but id has a strict type, we still want isStrictId id to be True.

isExportedId :: Var -> Bool

isExportedIdVar means "don't throw this away"

isLocalId :: Var -> Bool

isGlobalId :: Var -> Bool

isRecordSelector :: Id -> Bool

isNaughtyRecordSelector :: Id -> Bool

isPatSynRecordSelector :: Id -> Bool

isDataConRecordSelector :: Id -> Bool

isClassOpId :: Id -> Bool

isClassOpId_maybe :: Id -> Maybe Class

isDFunId :: Id -> Bool

isPrimOpId :: Id -> Bool

isPrimOpId_maybe :: Id -> Maybe PrimOp

isFCallId :: Id -> Bool

isFCallId_maybe :: Id -> Maybe ForeignCall

isDataConWorkId :: Id -> Bool

isDataConWorkId_maybe :: Id -> Maybe DataCon

isDataConWrapId :: Id -> Bool

isDataConWrapId_maybe :: Id -> Maybe DataCon

isDataConId_maybe :: Id -> Maybe DataCon

idDataCon :: Id -> DataCon

Get from either the worker or the wrapper Id to the DataCon. Currently used only in the desugarer.

INVARIANT: idDataCon (dataConWrapId d) = d: remember, dataConWrapId can return either the wrapper or the worker

isConLikeId :: Id -> Bool

isDeadEndId :: Var -> Bool

Returns true if an application to n args diverges or throws an exception See Note [Dead ends] in GHC.Types.Demand.

idIsFrom :: Module -> Id -> Bool

hasNoBinding :: Id -> Bool

Returns True of an Id which may not have a binding, even though it is defined in this module.

Join variables

type JoinId = Id

isJoinId :: Var -> Bool

isJoinId_maybe :: Var -> Maybe JoinArity

idJoinArity :: JoinId -> JoinArity

asJoinId :: Id -> JoinArity -> JoinId infixl 1

asJoinId_maybe :: Id -> Maybe JoinArity -> Id infixl 1

zapJoinId :: Id -> Id

Inline pragma stuff

idInlinePragma :: Id -> InlinePragma

setInlinePragma :: Id -> InlinePragma -> Id infixl 1

modifyInlinePragma :: Id -> (InlinePragma -> InlinePragma) -> Id

idInlineActivation :: Id -> Activation

setInlineActivation :: Id -> Activation -> Id infixl 1

idRuleMatchInfo :: Id -> RuleMatchInfo

One-shot lambdas

isOneShotBndr :: Var -> Bool

Returns whether the lambda associated with the Id is certainly applied at most once This one is the "business end", called externally. It works on type variables as well as Ids, returning True Its main purpose is to encapsulate the Horrible State Hack See Note [The state-transformer hack] in GHC.Core.Opt.Arity

isProbablyOneShotLambda :: Id -> Bool

setOneShotLambda :: Id -> Id

clearOneShotLambda :: Id -> Id

updOneShotInfo :: Id -> OneShotInfo -> Id

setIdOneShotInfo :: Id -> OneShotInfo -> Id infixl 1

isStateHackType :: Type -> Bool

stateHackOneShot :: OneShotInfo

Should we apply the state hack to values of this Type?

typeOneShot :: Type -> OneShotInfo

Reading IdInfo fields

idArity :: Id -> Arity

idCallArity :: Id -> Arity

idFunRepArity :: Id -> RepArity

idUnfolding :: Id -> Unfolding

realIdUnfolding :: Id -> Unfolding

idSpecialisation :: Id -> RuleInfo

idCoreRules :: Id -> [CoreRule]

idHasRules :: Id -> Bool

idCafInfo :: Id -> CafInfo infixl 1

idLFInfo_maybe :: Id -> Maybe LambdaFormInfo

idOneShotInfo :: Id -> OneShotInfo

idStateHackOneShotInfo :: Id -> OneShotInfo

Like idOneShotInfo, but taking the Horrible State Hack in to account See Note [The state-transformer hack] in GHC.Core.Opt.Arity

idOccInfo :: Id -> OccInfo

isNeverLevPolyId :: Id -> Bool

Writing IdInfo fields

setIdUnfolding :: Id -> Unfolding -> Id infixl 1

setCaseBndrEvald :: StrictnessMark -> Id -> Id

setIdArity :: Id -> Arity -> Id infixl 1

setIdCallArity :: Id -> Arity -> Id infixl 1

setIdSpecialisation :: Id -> RuleInfo -> Id infixl 1

setIdCafInfo :: Id -> CafInfo -> Id

setIdOccInfo :: Id -> OccInfo -> Id infixl 1

zapIdOccInfo :: Id -> Id

setIdLFInfo :: Id -> LambdaFormInfo -> Id

setIdDemandInfo :: Id -> Demand -> Id infixl 1

setIdStrictness :: Id -> StrictSig -> Id infixl 1

setIdCprInfo :: Id -> CprSig -> Id infixl 1

idDemandInfo :: Id -> Demand

idStrictness :: Id -> StrictSig

Accesses the Id's strictnessInfo.

idCprInfo :: Id -> CprSig