GHC.Types.Name

Description

GHC uses several kinds of name internally:

OccName: see GHC.Types.Name.Occurrence

RdrName: see GHC.Types.Name.Reader

Name is the type of names that have had their scoping and binding resolved. They have an OccName but also a Unique that disambiguates Names that have the same OccName and indeed is used for all Name comparison. Names also contain information about where they originated from, see GHC.Types.Name

Id: see GHC.Types.Id

Var: see GHC.Types.Var

Names are one of:

External, if they name things declared in other modules. Some external Names are wired in, i.e. they name primitives defined in the compiler itself

Internal, if they name things in the module being compiled. Some internal Names are system names, if they are names manufactured by the compiler

The main types

data Name

A unique, unambiguous name for something, containing information about where that thing originated.

data BuiltInSyntax

BuiltInSyntax is for things like (:), [] and tuples, which have special syntactic forms. They aren't in scope as such.

Constructors

BuiltInSyntax

UserSyntax

Creating Names

mkSystemName :: Unique -> OccName -> Name

Create a name brought into being by the compiler

mkSystemNameAt :: Unique -> OccName -> SrcSpan -> Name

mkInternalName :: Unique -> OccName -> SrcSpan -> Name

Create a name which is (for now at least) local to the current module and hence does not need a Module to disambiguate it from other Names

mkClonedInternalName :: Unique -> Name -> Name

mkDerivedInternalName :: (OccName -> OccName) -> Unique -> Name -> Name

mkSystemVarName :: Unique -> FastString -> Name

mkSysTvName :: Unique -> FastString -> Name

mkFCallName :: Unique -> String -> Name

Make a name for a foreign call

mkExternalName :: Unique -> Module -> OccName -> SrcSpan -> Name

Create a name which definitely originates in the given module

mkWiredInName :: Module -> OccName -> Unique -> TyThing -> BuiltInSyntax -> Name

Create a name which is actually defined by the compiler itself

Manipulating and deconstructing Names

nameUnique :: Name -> Unique

setNameUnique :: Name -> Unique -> Name

nameOccName :: Name -> OccName

nameNameSpace :: Name -> NameSpace

nameModule :: HasDebugCallStack => Name -> Module

nameModule_maybe :: Name -> Maybe Module

setNameLoc :: Name -> SrcSpan -> Name

tidyNameOcc :: Name -> OccName -> Name

localiseName :: Name -> Name

Make the Name into an internal name, regardless of what it was to begin with

nameSrcLoc :: Name -> SrcLoc

nameSrcSpan :: Name -> SrcSpan

pprNameDefnLoc :: Name -> SDoc

pprDefinedAt :: Name -> SDoc

Predicates on Names

isSystemName :: Name -> Bool

isInternalName :: Name -> Bool

isExternalName :: Name -> Bool

isTyVarName :: Name -> Bool

isTyConName :: Name -> Bool

isDataConName :: Name -> Bool

isValName :: Name -> Bool

isVarName :: Name -> Bool

isDynLinkName :: Platform -> Module -> Name -> Bool

Will the Name come from a dynamically linked package?

isWiredInName :: Name -> Bool

isWiredIn :: NamedThing thing => thing -> Bool

isBuiltInSyntax :: Name -> Bool

isHoleName :: Name -> Bool

wiredInNameTyThing_maybe :: Name -> Maybe TyThing

nameIsLocalOrFrom :: Module -> Name -> Bool

Returns True if the name is (a) Internal (b) External but from the specified module (c) External but from the interactive package

The key idea is that False means: the entity is defined in some other module you can find the details (type, fixity, instances) in some interface file those details will be stored in the EPT or HPT

True means: the entity is defined in this module or earlier in the GHCi session you can find details (type, fixity, instances) in the TcGblEnv or TcLclEnv

The isInteractiveModule part is because successive interactions of a GHCi session each give rise to a fresh module (Ghci1, Ghci2, etc), but they all come from the magic interactive package; and all the details are kept in the TcLclEnv, TcGblEnv, NOT in the HPT or EPT. See Note [The interactive package] in GHC.Runtime.Context

nameIsHomePackage :: Module -> Name -> Bool

nameIsHomePackageImport :: Module -> Name -> Bool

nameIsFromExternalPackage :: HomeUnit -> Name -> Bool

Returns True if the Name comes from some other package: neither this package nor the interactive package.

stableNameCmp :: Name -> Name -> Ordering

Compare Names lexicographically This only works for Names that originate in the source code or have been tidied.

Class NamedThing and overloaded friends

class NamedThing a where

A class allowing convenient access to the Name of various datatypes

Minimal complete definition

getName

Methods

getOccName :: a -> OccName

getName :: a -> Name

getSrcLoc :: NamedThing a => a -> SrcLoc

getSrcSpan :: NamedThing a => a -> SrcSpan

getOccString :: NamedThing a => a -> String

getOccFS :: NamedThing a => a -> FastString

pprInfixName :: (Outputable a, NamedThing a) => a -> SDoc

pprPrefixName :: NamedThing a => a -> SDoc

pprModulePrefix :: PprStyle -> Module -> OccName -> SDoc

pprNameUnqualified :: Name -> SDoc

Print the string of Name unqualifiedly directly.

nameStableString :: Name -> String

Get a string representation of a Name that's unique and stable across recompilations. Used for deterministic generation of binds for derived instances. eg. "$aeson_70dylHtv1FFGeai1IoxcQr$Data.Aeson.Types.Internal$String"

module GHC.Types.Name.Occurrence