GHC.Types.Name.Ppr

Documentation

mkPrintUnqualified :: UnitEnv -> GlobalRdrEnv -> PrintUnqualified

Creates some functions that work out the best ways to format names for the user according to a set of heuristics.

mkQualModule :: UnitState -> HomeUnit -> QueryQualifyModule

Creates a function for formatting modules based on two heuristics: (1) if the module is the current module, don't qualify, and (2) if there is only one exposed package which exports this module, don't qualify.

mkQualPackage :: UnitState -> QueryQualifyPackage

Creates a function for formatting packages based on two heuristics: (1) don't qualify if the package in question is "main", and (2) only qualify with a unit id if the package ID would be ambiguous.

pkgQual :: UnitState -> PrintUnqualified

A function which only qualifies package names if necessary; but qualifies all other identifiers.