GHC.Types.SafeHaskell

Description

This stuff here is related to supporting the Safe Haskell extension, primarily about storing under what trust type a module has been compiled.

Documentation

type IsSafeImport = Bool

Is an import a safe import?

data SafeHaskellMode

The various Safe Haskell modes

Constructors

Sf_None
inferred unsafe

Sf_Unsafe
declared and checked

Sf_Trustworthy
declared and checked

Sf_Safe
declared and checked

Sf_SafeInferred
inferred as safe

Sf_Ignore
-fno-safe-haskell state

data IfaceTrustInfo

Safe Haskell information for ModIface Simply a wrapper around SafeHaskellMode to sepperate iface and flags

getSafeMode :: IfaceTrustInfo -> SafeHaskellMode

setSafeMode :: SafeHaskellMode -> IfaceTrustInfo

noIfaceTrustInfo :: IfaceTrustInfo