GHC.Types.SourceFile

Documentation

data HscSource

Constructors

HsSrcFile
.hs file

HsBootFile
.hs-boot file

HsigFile
.hsig file

data SourceModified

Indicates whether a given module's source has been modified since it was last compiled.

Constructors

SourceModified
the source has been modified

SourceUnmodified
the source has not been modified. Compilation may or may not be necessary, depending on whether any dependencies have changed since we last compiled.

SourceUnmodifiedAndStable
the source has not been modified, and furthermore all of its (transitive) dependencies are up to date; it definitely does not need to be recompiled. This is important for two reasons: (a) we can omit the version check in checkOldIface, and (b) if the module used TH splices we don't need to force recompilation.

isHsBootOrSig :: HscSource -> Bool

isHsigFile :: HscSource -> Bool

hscSourceString :: HscSource -> String