GHC.Types.SourceText

Description

Source text

Keeping Source Text for source to source conversions

Documentation

data SourceText

Constructors

SourceText String

NoSourceText
For when code is generated, e.g. TH, deriving. The pretty printer will then make its own representation of the item.

pprWithSourceText :: SourceText -> SDoc -> SDoc

Special combinator for showing string literals.

Literals

data IntegralLit

Integral Literal

Used (instead of Integer) to represent negative zegative zero which is required for NegativeLiterals extension to correctly parse `-0::Double` as negative zero. See also #13211.

Constructors

IL

Fields

il_text :: SourceText

il_neg :: Bool

il_value :: Integer

data FractionalLit

Fractional Literal

Used (instead of Rational) to represent exactly the floating point literal that we encountered in the user's source program. This allows us to pretty-print exactly what the user wrote, which is important e.g. for floating point numbers that can't represented as Doubles (we used to via Double for pretty-printing). See also #2245. Note [FractionalLit representation] in GHC.HsToCore.Match.Literal The actual value then is: sign * fl_signi * (fl_exp_base^fl_exp) where sign = if fl_neg then (-1) else 1

For example FL { fl_neg = True, fl_signi = 5.3, fl_exp = 4, fl_exp_base = Base10 } denotes -5300

Constructors

FL

Fields

fl_text :: SourceText

How the value was written in the source

fl_neg :: Bool

fl_signi :: Rational

fl_exp :: Integer

fl_exp_base :: FractionalExponentBase

data StringLiteral

A String Literal in the source, including its original raw format for use by source to source manipulation tools.

Constructors

StringLiteral

Fields

sl_st :: SourceText

sl_fs :: FastString

sl_tc :: Maybe RealSrcSpan

negateIntegralLit :: IntegralLit -> IntegralLit

negateFractionalLit :: FractionalLit -> FractionalLit

mkIntegralLit :: Integral a => a -> IntegralLit

mkTHFractionalLit :: Rational -> FractionalLit

rationalFromFractionalLit :: FractionalLit -> Rational

integralFractionalLit :: Bool -> Integer -> FractionalLit

The integer should already be negated if it's negative.

mkSourceFractionalLit :: String -> Bool -> Integer -> Integer -> FractionalExponentBase -> FractionalLit

The arguments should already be negated if they are negative.

data FractionalExponentBase

Constructors

Base2

Base10

fractionalLitFromRational :: Rational -> FractionalLit

mkFractionalLit :: SourceText -> Bool -> Rational -> Integer -> FractionalExponentBase -> FractionalLit