GHC.Types.TyThing.Ppr

Documentation

pprTyThing :: ShowSub -> TyThing -> SDoc

Pretty-prints a TyThing.

pprTyThingInContext :: ShowSub -> TyThing -> SDoc

Pretty-prints a TyThing in context: that is, if the entity is a data constructor, record selector, or class method, then the entity's parent declaration is pretty-printed with irrelevant parts omitted.

pprTyThingLoc :: TyThing -> SDoc

Pretty-prints a TyThing with its defining location.

pprTyThingInContextLoc :: TyThing -> SDoc

Like pprTyThingInContext, but adds the defining location.

pprTyThingHdr :: TyThing -> SDoc

Pretty-prints the TyThing header. For functions and data constructors the function is equivalent to pprTyThing but for type constructors and classes it prints only the header part of the declaration.

pprTypeForUser :: Type -> SDoc

pprFamInst :: FamInst -> SDoc

Pretty-prints a FamInst (type/data family instance) with its defining location.