GHC.Types.Unique.DFM

Unique-keyed deterministic mappings

data UniqDFM key ele

Type of unique deterministic finite maps

The key is just here to keep us honest. It's always safe to use a single type as key. If two types don't overlap in their uniques it's also safe to index the same map at multiple key types. But this is very much discouraged.

Manipulating those mappings

emptyUDFM :: UniqDFM key elt

unitUDFM :: Uniquable key => key -> elt -> UniqDFM key elt

addToUDFM :: Uniquable key => UniqDFM key elt -> key -> elt -> UniqDFM key elt

addToUDFM_C :: Uniquable key => (elt -> elt -> elt) -> UniqDFM key elt -> key -> elt -> UniqDFM key elt

addToUDFM_C_Directly :: (elt -> elt -> elt) -> UniqDFM key elt -> Unique -> elt -> UniqDFM key elt

addToUDFM_Directly :: UniqDFM key elt -> Unique -> elt -> UniqDFM key elt

addListToUDFM :: Uniquable key => UniqDFM key elt -> [(key, elt)] -> UniqDFM key elt

delFromUDFM :: Uniquable key => UniqDFM key elt -> key -> UniqDFM key elt

delListFromUDFM :: Uniquable key => UniqDFM key elt -> [key] -> UniqDFM key elt

Delete a list of elements from a UniqDFM

adjustUDFM :: Uniquable key => (elt -> elt) -> UniqDFM key elt -> key -> UniqDFM key elt

Apply a function to a particular element

adjustUDFM_Directly :: (elt -> elt) -> UniqDFM key elt -> Unique -> UniqDFM key elt

Apply a function to a particular element

alterUDFM :: Uniquable key => (Maybe elt -> Maybe elt) -> UniqDFM key elt -> key -> UniqDFM key elt

The expression (alterUDFM f k map) alters value x at k, or absence thereof. alterUDFM can be used to insert, delete, or update a value in UniqDFM. Use addToUDFM, delFromUDFM or adjustUDFM when possible, they are more efficient.

mapUDFM :: (elt1 -> elt2) -> UniqDFM key elt1 -> UniqDFM key elt2

Map a function over every value in a UniqDFM

mapMaybeUDFM :: forall elt1 elt2 key. (elt1 -> Maybe elt2) -> UniqDFM key elt1 -> UniqDFM key elt2

plusUDFM :: UniqDFM key elt -> UniqDFM key elt -> UniqDFM key elt

plusUDFM_C :: (elt -> elt -> elt) -> UniqDFM key elt -> UniqDFM key elt -> UniqDFM key elt

lookupUDFM :: Uniquable key => UniqDFM key elt -> key -> Maybe elt

lookupUDFM_Directly :: UniqDFM key elt -> Unique -> Maybe elt

elemUDFM :: Uniquable key => key -> UniqDFM key elt -> Bool

foldUDFM :: (elt -> a -> a) -> a -> UniqDFM key elt -> a

Performs a deterministic fold over the UniqDFM. It's O(n log n) while the corresponding function on UniqFM is O(n).

eltsUDFM :: UniqDFM key elt -> [elt]

filterUDFM :: (elt -> Bool) -> UniqDFM key elt -> UniqDFM key elt

filterUDFM_Directly :: (Unique -> elt -> Bool) -> UniqDFM key elt -> UniqDFM key elt

isNullUDFM :: UniqDFM key elt -> Bool

sizeUDFM :: UniqDFM key elt -> Int

intersectUDFM :: UniqDFM key elt -> UniqDFM key elt -> UniqDFM key elt

udfmIntersectUFM :: UniqDFM key elt1 -> UniqFM key elt2 -> UniqDFM key elt1

disjointUDFM :: UniqDFM key elt -> UniqDFM key elt -> Bool

disjointUdfmUfm :: UniqDFM key elt -> UniqFM key elt2 -> Bool

equalKeysUDFM :: UniqDFM key a -> UniqDFM key b -> Bool

minusUDFM :: UniqDFM key elt1 -> UniqDFM key elt2 -> UniqDFM key elt1

listToUDFM :: Uniquable key => [(key, elt)] -> UniqDFM key elt

listToUDFM_Directly :: [(Unique, elt)] -> UniqDFM key elt

udfmMinusUFM :: UniqDFM key elt1 -> UniqFM key elt2 -> UniqDFM key elt1

ufmMinusUDFM :: UniqFM key elt1 -> UniqDFM key elt2 -> UniqFM key elt1

partitionUDFM :: (elt -> Bool) -> UniqDFM key elt -> (UniqDFM key elt, UniqDFM key elt)

Partition UniqDFM into two UniqDFMs according to the predicate

anyUDFM :: (elt -> Bool) -> UniqDFM key elt -> Bool

allUDFM :: (elt -> Bool) -> UniqDFM key elt -> Bool

pprUniqDFM :: (a -> SDoc) -> UniqDFM key a -> SDoc

pprUDFM

Arguments

:: UniqDFM key a
The things to be pretty printed

-> ([a] -> SDoc)
The pretty printing function to use on the elements

-> SDoc
SDoc where the things have been pretty printed

udfmToList :: UniqDFM key elt -> [(Unique, elt)]

Converts UniqDFM to a list, with elements in deterministic order. It's O(n log n) while the corresponding function on UniqFM is O(n).

udfmToUfm :: UniqDFM key elt -> UniqFM key elt

This allows for lossy conversion from UniqDFM to UniqFM

nonDetStrictFoldUDFM :: (elt -> a -> a) -> a -> UniqDFM key elt -> a

Performs a nondeterministic strict fold over the UniqDFM. It's O(n), same as the corresponding function on UniqFM. If you use this please provide a justification why it doesn't introduce nondeterminism.

unsafeCastUDFMKey :: UniqDFM key1 elt -> UniqDFM key2 elt

Cast the key domain of a UniqFM.

As long as the domains don't overlap in their uniques this is safe.

alwaysUnsafeUfmToUdfm :: UniqFM key elt -> UniqDFM key elt