GHC.Types.Unique.Set

Unique set type

data UniqSet a

Set of Uniquable values

getUniqSet :: UniqSet a -> UniqFM a a

pprUniqSet :: (a -> SDoc) -> UniqSet a -> SDoc

Manipulating these sets

emptyUniqSet :: UniqSet a

unitUniqSet :: Uniquable a => a -> UniqSet a

mkUniqSet :: Uniquable a => [a] -> UniqSet a

addOneToUniqSet :: Uniquable a => UniqSet a -> a -> UniqSet a

addListToUniqSet :: Uniquable a => UniqSet a -> [a] -> UniqSet a

delOneFromUniqSet :: Uniquable a => UniqSet a -> a -> UniqSet a

delOneFromUniqSet_Directly :: UniqSet a -> Unique -> UniqSet a

delListFromUniqSet :: Uniquable a => UniqSet a -> [a] -> UniqSet a

delListFromUniqSet_Directly :: UniqSet a -> [Unique] -> UniqSet a

unionUniqSets :: UniqSet a -> UniqSet a -> UniqSet a

unionManyUniqSets :: [UniqSet a] -> UniqSet a

minusUniqSet :: UniqSet a -> UniqSet a -> UniqSet a

uniqSetMinusUFM :: UniqSet key -> UniqFM key b -> UniqSet key

uniqSetMinusUDFM :: UniqSet key -> UniqDFM key b -> UniqSet key

intersectUniqSets :: UniqSet a -> UniqSet a -> UniqSet a

disjointUniqSets :: UniqSet a -> UniqSet a -> Bool

restrictUniqSetToUFM :: UniqSet key -> UniqFM key b -> UniqSet key

uniqSetAny :: (a -> Bool) -> UniqSet a -> Bool

uniqSetAll :: (a -> Bool) -> UniqSet a -> Bool

elementOfUniqSet :: Uniquable a => a -> UniqSet a -> Bool

elemUniqSet_Directly :: Unique -> UniqSet a -> Bool

filterUniqSet :: (a -> Bool) -> UniqSet a -> UniqSet a

filterUniqSet_Directly :: (Unique -> elt -> Bool) -> UniqSet elt -> UniqSet elt

sizeUniqSet :: UniqSet a -> Int

isEmptyUniqSet :: UniqSet a -> Bool

lookupUniqSet :: Uniquable key => UniqSet key -> key -> Maybe key

What's the point you might ask? We might have changed an object without it's key changing. In which case this lookup makes sense.

lookupUniqSet_Directly :: UniqSet a -> Unique -> Maybe a

partitionUniqSet :: (a -> Bool) -> UniqSet a -> (UniqSet a, UniqSet a)

mapUniqSet :: Uniquable b => (a -> b) -> UniqSet a -> UniqSet b

unsafeUFMToUniqSet :: UniqFM a a -> UniqSet a

unsafeUFMToUniqSet converts a UniqFM a into a UniqSet a assuming, without checking, that it maps each Unique to a value that has that Unique. See Note [UniqSet invariant].

nonDetEltsUniqSet :: UniqSet elt -> [elt]

nonDetKeysUniqSet :: UniqSet elt -> [Unique]

nonDetStrictFoldUniqSet :: (elt -> a -> a) -> a -> UniqSet elt -> a

mapMaybeUniqSet_sameUnique :: (a -> Maybe b) -> UniqSet a -> UniqSet b

Like mapMaybe, but you must ensure the passed in function does not change the Unique.

newtype UniqueSet

Set of Unique values

Similar to 'UniqSet Unique' but with a more compact representation.

Constructors

US

Fields

unUniqueSet :: Word64Set

nullUniqueSet :: UniqueSet -> Bool

sizeUniqueSet :: UniqueSet -> Int

memberUniqueSet :: Unique -> UniqueSet -> Bool

emptyUniqueSet :: UniqueSet

singletonUniqueSet :: Unique -> UniqueSet

insertUniqueSet :: Unique -> UniqueSet -> UniqueSet

deleteUniqueSet :: Unique -> UniqueSet -> UniqueSet

differenceUniqueSet :: UniqueSet -> UniqueSet -> UniqueSet

unionUniqueSet :: UniqueSet -> UniqueSet -> UniqueSet

unionsUniqueSet :: [UniqueSet] -> UniqueSet

intersectionUniqueSet :: UniqueSet -> UniqueSet -> UniqueSet

isSubsetOfUniqueSet :: UniqueSet -> UniqueSet -> Bool

filterUniqueSet :: (Unique -> Bool) -> UniqueSet -> UniqueSet

foldlUniqueSet :: (a -> Unique -> a) -> a -> UniqueSet -> a

foldrUniqueSet :: (Unique -> b -> b) -> b -> UniqueSet -> b

elemsUniqueSet :: UniqueSet -> [Unique]

fromListUniqueSet :: [Unique] -> UniqueSet