GHC.Types.Unique.Supply

Main data type

data UniqSupply

Unique Supply

A value of type UniqSupply is unique, and it can supply one distinct Unique. Also, from the supply, one can also manufacture an arbitrary number of further UniqueSupply values, which will be distinct from the first and from all others.

Operations on supplies

uniqFromSupply :: UniqSupply -> Unique

Obtain the Unique from this particular UniqSupply

uniqsFromSupply :: UniqSupply -> [Unique]

Obtain an infinite list of Unique that can be generated by constant splitting of the supply

takeUniqFromSupply :: UniqSupply -> (Unique, UniqSupply)

Obtain the Unique from this particular UniqSupply, and a new supply

uniqFromMask :: Char -> IO Unique

mkSplitUniqSupply :: Char -> IO UniqSupply

Create a unique supply out of thin air. The "mask" (Char) supplied is purely cosmetic, making it easier to figure out where a Unique was born. See Note [Uniques and masks].

The payload part of the Uniques allocated from this UniqSupply are guaranteed distinct wrt all other supplies, regardless of their "mask". This is achieved by allocating the payload part from a single source of Uniques, namely genSym, shared across all UniqSupply's.

splitUniqSupply :: UniqSupply -> (UniqSupply, UniqSupply)

Build two UniqSupply from a single one, each of which can supply its own Unique.

listSplitUniqSupply :: UniqSupply -> [UniqSupply]

Create an infinite list of UniqSupply from a single one

Unique supply monad and its abstraction

data UniqSM result

A monad which just gives the ability to obtain Uniques

class Monad m => MonadUnique m where

A monad for generating unique identifiers

Minimal complete definition

getUniqueSupplyM

Methods

getUniqueSupplyM :: m UniqSupply

Get a new UniqueSupply

getUniqueM :: m Unique

Get a new unique identifier

getUniquesM :: m [Unique]

Get an infinite list of new unique identifiers

Operations on the monad

initUs :: UniqSupply -> UniqSM a -> (a, UniqSupply)

Run the UniqSM action, returning the final UniqSupply

initUs_ :: UniqSupply -> UniqSM a -> a

Run the UniqSM action, discarding the final UniqSupply

Set supply strategy

initUniqSupply :: Word -> Int -> IO ()