GHC.Types.Var

Description

GHC uses several kinds of name internally:

OccName: see GHC.Types.Name.Occurrence

RdrName: see GHC.Types.Name.Reader

Name: see GHC.Types.Name

Id: see GHC.Types.Id

Var is a synonym for the Id type but it may additionally potentially contain type variables, which have a Kind rather than a Type and only contain some extra details during typechecking.

These Var names may either be global or local, see GHC.Types.Var

Global Vars and Vars are those that are imported or correspond to a data constructor, primitive operation, or record selectors. Local Vars and Vars are those bound within an expression (e.g. by a lambda) or at the top level of the module being compiled.

The main data type and synonyms

data Var

Variable

Essentially a typed Name, that may also contain some additional information about the Var and its use sites.

type CoVar = Id

Coercion Variable

type Id = Var

Identifier

type NcId = Id

type DictId = EvId

Dictionary Identifier

type DFunId = Id

Dictionary Function Identifier

type EvVar = EvId

Evidence Variable

type EqVar = EvId

Equality Variable

type EvId = Id

Evidence Identifier

type IpId = EvId

Implicit parameter Identifier

type JoinId = Id

type TyVar = Var

Type or kind Variable

type TcTyVar = Var

Type variable that might be a metavariable

type TypeVar = Var

Type Variable

type KindVar = Var

Kind Variable

type TKVar = Var

Type or Kind Variable

type TyCoVar = Id

Type or Coercion Variable

In and Out variants

type InVar = Var

type InCoVar = CoVar

type InId = Id

type InTyVar = TyVar

type OutVar = Var

type OutCoVar = CoVar

type OutId = Id

type OutTyVar = TyVar

Taking Vars apart

varName :: Var -> Name

varUnique :: Var -> Unique

varType :: Var -> Kind

The type or kind of the Var in question

varMult :: Var -> Mult

varMultMaybe :: Id -> Maybe Mult

Modifying Vars

setVarName :: Var -> Name -> Var

setVarUnique :: Var -> Unique -> Var

setVarType :: Var -> Type -> Var

updateVarType :: (Type -> Type) -> Var -> Var

Update a Vars type. Does not update the multiplicity stored in an Var, if any. Because of the possibility for abuse, ASSERTs that there is no multiplicity to update.

updateVarTypeM :: Monad m => (Type -> m Type) -> Var -> m Var

Update a Vars type monadically. Does not update the multiplicity stored in an Var, if any. Because of the possibility for abuse, ASSERTs that there is no multiplicity to update.

Constructing, taking apart, modifying Vars

mkGlobalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

mkLocalVar :: IdDetails -> Name -> Mult -> Type -> IdInfo -> Id

mkExportedLocalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

Exported Vars will not be removed as dead code

mkCoVar :: Name -> Type -> CoVar

idInfo :: HasDebugCallStack => Id -> IdInfo

idDetails :: Id -> IdDetails

lazySetIdInfo :: Id -> IdInfo -> Var

setIdDetails :: Id -> IdDetails -> Id

globaliseId :: Id -> Id

If it's a local, make it global

setIdExported :: Id -> Id

Exports the given local Var. Can also be called on global Vars, such as data constructors and class operations, which are born as global Vars and automatically exported

setIdNotExported :: Id -> Id

We can only do this to LocalIds

setIdMult :: Id -> Mult -> Id

updateIdTypeButNotMult :: (Type -> Type) -> Id -> Id

updateIdTypeAndMult :: (Type -> Type) -> Id -> Id

updateIdTypeAndMultM :: Monad m => (Type -> m Type) -> Id -> m Id

Predicates

isId :: Var -> Bool

Is this a value-level (i.e., computationally relevant) Varentifier? Satisfies isId = not . isTyVar.

isTyVar :: Var -> Bool

Is this a type-level (i.e., computationally irrelevant, thus erasable) variable? Satisfies isTyVar = not . isId.

isTcTyVar :: Var -> Bool

isLocalVar :: Var -> Bool

isLocalVar returns True for type variables as well as local Vars These are the variables that we need to pay attention to when finding free variables, or doing dependency analysis.

isLocalId :: Var -> Bool

isCoVar :: Var -> Bool

Is this a coercion variable? Satisfies isId v ==> isCoVar v == not (isNonCoVarId v).

isNonCoVarId :: Var -> Bool

Is this a term variable (Var) that is not a coercion variable? Satisfies isId v ==> isCoVar v == not (isNonCoVarId v).

isTyCoVar :: Var -> Bool

isGlobalId :: Var -> Bool

isExportedId :: Var -> Bool

isExportedIdVar means "don't throw this away"

mustHaveLocalBinding :: Var -> Bool

mustHaveLocalBinding returns True of Vars and Vars that must have a binding in this module. The converse is not quite right: there are some global Vars that must have bindings, such as record selectors. But that doesn't matter, because it's only used for assertions

ArgFlags

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, TyCoVarBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible Specificity

Required

Bundled Patterns

pattern Specified :: ArgFlag

pattern Inferred :: ArgFlag

data AnonArgFlag

The non-dependent version of ArgFlag. See Note [AnonArgFlag] Appears here partly so that it's together with its friends ArgFlag and ForallVisFlag, but also because it is used in IfaceType, rather early in the compilation chain

Constructors

VisArg
Used for (->): an ordinary non-dependent arrow. The argument is visible in source code.

InvisArg
Used for (=>): a non-dependent predicate arrow. The argument is invisible in source code.

data Specificity

Whether an Invisible argument may appear in source Haskell.

Constructors

InferredSpec
the argument may not appear in source Haskell, it is only inferred.

SpecifiedSpec
the argument may appear in source Haskell, but isn't required.

isVisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is written in Haskell?

isInvisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is not written in Haskell?

isInferredArgFlag :: ArgFlag -> Bool

sameVis :: ArgFlag -> ArgFlag -> Bool

Do these denote the same level of visibility? Required arguments are visible, others are not. So this function equates Specified and Inferred. Used for printing.

TyVar's

data VarBndr var argf

Constructors

Bndr var argf

type TyCoVarBinder = VarBndr TyCoVar ArgFlag

Variable Binder

A TyCoVarBinder is the binder of a ForAllTy It's convenient to define this synonym here rather its natural home in GHC.Core.TyCo.Rep, because it's used in GHC.Core.DataCon.hs-boot

A TyVarBinder is a binder with only TyVar

type TyVarBinder = VarBndr TyVar ArgFlag

type InvisTVBinder = VarBndr TyVar Specificity

type ReqTVBinder = VarBndr TyVar ()

binderVar :: VarBndr tv argf -> tv

binderVars :: [VarBndr tv argf] -> [tv]

binderArgFlag :: VarBndr tv argf -> argf

binderType :: VarBndr TyCoVar argf -> Type

mkTyCoVarBinder :: vis -> TyCoVar -> VarBndr TyCoVar vis

Make a named binder

mkTyCoVarBinders :: vis -> [TyCoVar] -> [VarBndr TyCoVar vis]

Make many named binders

mkTyVarBinder :: vis -> TyVar -> VarBndr TyVar vis

Make a named binder var should be a type variable

mkTyVarBinders :: vis -> [TyVar] -> [VarBndr TyVar vis]

Make many named binders Input vars should be type variables

isTyVarBinder :: TyCoVarBinder -> Bool

tyVarSpecToBinder :: VarBndr a Specificity -> VarBndr a ArgFlag

tyVarSpecToBinders :: [VarBndr a Specificity] -> [VarBndr a ArgFlag]

tyVarReqToBinder :: VarBndr a () -> VarBndr a ArgFlag

tyVarReqToBinders :: [VarBndr a ()] -> [VarBndr a ArgFlag]

mapVarBndr :: (var -> var') -> VarBndr var flag -> VarBndr var' flag

mapVarBndrs :: (var -> var') -> [VarBndr var flag] -> [VarBndr var' flag]

lookupVarBndr :: Eq var => var -> [VarBndr var flag] -> Maybe flag

Constructing TyVar's

mkTyVar :: Name -> Kind -> TyVar

mkTcTyVar :: Name -> Kind -> TcTyVarDetails -> TyVar

Taking Vars apart

tyVarName :: TyVar -> Name

tyVarKind :: TyVar -> Kind

tcTyVarDetails :: TyVar -> TcTyVarDetails

setTcTyVarDetails :: TyVar -> TcTyVarDetails -> TyVar

Modifying Vars

setTyVarName :: TyVar -> Name -> TyVar

setTyVarUnique :: TyVar -> Unique -> TyVar

setTyVarKind :: TyVar -> Kind -> TyVar

updateTyVarKind :: (Kind -> Kind) -> TyVar -> TyVar

updateTyVarKindM :: Monad m => (Kind -> m Kind) -> TyVar -> m TyVar

nonDetCmpVar :: Var -> Var -> Ordering

Compare Vars by their Uniques. This is what Ord Var does, provided here to make it explicit at the call-site that it can introduce non-determinism. See Note [Unique Determinism]