GHC.Unit.Finder.Types

Documentation

data FinderCache

Constructors

FinderCache

Fields

fcModuleCache :: IORef FinderCacheState

fcFileCache :: IORef FileCacheState

type FinderCacheState = InstalledModuleEnv InstalledFindResult

The FinderCache maps modules to the result of searching for that module. It records the results of searching for modules along the search path. On :load, we flush the entire contents of this cache.

data FindResult

The result of searching for an imported module.

NB: FindResult manages both user source-import lookups (which can result in GenModule) as well as direct imports for interfaces (which always result in InstalledModule).

Constructors

Found ModLocation Module
The module was found

NoPackage Unit
The requested unit was not found

FoundMultiple [(Module, ModuleOrigin)]
_Error_: both in multiple packages

NotFound
Not found

Fields

fr_paths :: [FilePath]

Places where I looked

fr_pkg :: Maybe Unit

Just p => module is in this unit's manifest, but couldn't find the .hi file

fr_mods_hidden :: [Unit]

Module is in these units, but the *module* is hidden

fr_pkgs_hidden :: [Unit]

Module is in these units, but the *unit* is hidden

fr_unusables :: [UnusableUnit]

Module is in these units, but it is unusable

fr_suggestions :: [ModuleSuggestion]

Possible mis-spelled modules

data InstalledFindResult

Constructors

InstalledFound ModLocation InstalledModule

InstalledNoPackage UnitId

InstalledNotFound [FilePath] (Maybe UnitId)

data FinderOpts

Locations and information the finder cares about.

Should be taken from DynFlags via initFinderOpts.

Constructors

FinderOpts

Fields

finder_importPaths :: [FilePath]

Where are we allowed to look for Modules and Source files

finder_lookupHomeInterfaces :: Bool

When looking up a home module:

True: search interface files (e.g. in '-c' mode)

False: search source files (e.g. in '--make' mode)

finder_bypassHiFileCheck :: Bool

Don't check that an imported interface file actually exists if it can only be at one location. The interface will be reported as InstalledFound even if the file doesn't exist, so this is only useful in specific cases (e.g. to generate dependencies with `ghc -M`)

finder_ways :: Ways

finder_enableSuggestions :: Bool

If we encounter unknown modules, should we suggest modules that have a similar name.

finder_workingDirectory :: Maybe FilePath

finder_thisPackageName :: Maybe FastString

finder_hiddenModules :: Set ModuleName

finder_reexportedModules :: Set ModuleName

finder_hieDir :: Maybe FilePath

finder_hieSuf :: String

finder_hiDir :: Maybe FilePath

finder_hiSuf :: String

finder_dynHiSuf :: String

finder_objectDir :: Maybe FilePath

finder_objectSuf :: String

finder_dynObjectSuf :: String

finder_stubDir :: Maybe FilePath