GHC.Unit.Module

Documentation

module GHC.Unit.Types

The ModuleName type

module GHC.Unit.Module.Name

The ModLocation type

module GHC.Unit.Module.Location

ModuleEnv

module GHC.Unit.Module.Env

Generalization

getModuleInstantiation :: Module -> (InstalledModule, Maybe InstantiatedModule)

Given a possibly on-the-fly instantiated module, split it into a Module that we definitely can find on-disk, as well as an instantiation if we need to instantiate it on the fly. If the instantiation is Nothing no on-the-fly renaming is needed.

getUnitInstantiations :: Unit -> (UnitId, Maybe InstantiatedUnit)

Return the unit-id this unit is an instance of and the module instantiations (if any).

uninstantiateInstantiatedUnit :: InstantiatedUnit -> InstantiatedUnit

Remove instantiations of the given instantiated unit

uninstantiateInstantiatedModule :: InstantiatedModule -> InstantiatedModule

Remove instantiations of the given module instantiated unit

The Module type

mkHoleModule :: ModuleName -> GenModule (GenUnit u)

Create a hole Module

isHoleModule :: GenModule (GenUnit u) -> Bool

Test if a Module is not instantiated

stableModuleCmp :: Module -> Module -> Ordering

This gives a stable ordering, as opposed to the Ord instance which gives an ordering based on the Uniques of the components, which may not be stable from run to run of the compiler.

moduleStableString :: Module -> String

Get a string representation of a Module that's unique and stable across recompilations. eg. "$aeson_70dylHtv1FFGeai1IoxcQr$Data.Aeson.Types.Internal"

moduleIsDefinite :: Module -> Bool

A Module is definite if it has no free holes.

class HasModule m where

Methods

getModule :: m Module

class ContainsModule t where

Methods

extractModule :: t -> Module

installedModuleEq :: InstalledModule -> Module -> Bool

Test if a Module corresponds to a given InstalledModule, modulo instantiation.