GHC.Unit.Module.ModDetails

Documentation

data ModDetails

The ModDetails is essentially a cache for information in the ModIface for home modules only. Information relating to packages will be loaded into global environments in ExternalPackageState.

Constructors

ModDetails

Fields

md_exports :: [AvailInfo]

md_types :: !TypeEnv

Local type environment for this particular module Includes Ids, TyCons, PatSyns

md_insts :: ![ClsInst]

DFunIds for the instances in this module

md_fam_insts :: ![FamInst]

md_rules :: ![CoreRule]

Domain may include Ids from other modules

md_anns :: ![Annotation]

Annotations present in this module: currently they only annotate things also declared in this module

md_complete_matches :: [CompleteMatch]

Complete match pragmas for this module

emptyModDetails :: ModDetails

Constructs an empty ModDetails