GHC.Unit.Module.WholeCoreBindings

Documentation

data WholeCoreBindings

Constructors

WholeCoreBindings

Fields

wcb_bindings :: [IfaceBindingX IfaceMaybeRhs IfaceTopBndrInfo]

serialised tidied core bindings.

wcb_module :: Module

The module which the bindings are for

wcb_mod_location :: ModLocation

The location where the sources reside. | Stubs for foreign declarations and files added via addForeignFilePath.

wcb_foreign :: IfaceForeign

newtype IfaceCLabel

Wrapper for avoiding a dependency on Binary and NFData in CLabel.

Constructors

IfaceCLabel CStubLabel

data IfaceCStubs

Simplified encoding of ForeignStubs for interface serialization.

See Note [Foreign stubs and TH bytecode linking]

Constructors

IfaceCStubs

Fields

header :: String

source :: String

initializers :: [IfaceCLabel]

finalizers :: [IfaceCLabel]

binary_put_ForeignSrcLang :: WriteBinHandle -> ForeignSrcLang -> IO ()

Binary put_ for ForeignSrcLang.

binary_get_ForeignSrcLang :: ReadBinHandle -> IO ForeignSrcLang

Binary get for ForeignSrcLang.

data IfaceForeignFile

A source file added from Template Haskell using qAddForeignFilePath, for storage in interfaces.

See Note [Foreign stubs and TH bytecode linking]

Constructors

IfaceForeignFile

Fields

lang :: ForeignSrcLang

The language is specified by the user.

source :: String

The contents of the file, which will be written to a temporary file when loaded from an interface.

extension :: FilePath

The extension used by the user is preserved, to avoid confusing external tools with an unexpected .c file or similar.

data IfaceForeign

Constructors

IfaceForeign

Fields

stubs :: Maybe IfaceCStubs

files :: [IfaceForeignFile]

emptyIfaceForeign :: IfaceForeign

encodeIfaceForeign :: Logger -> DynFlags -> ForeignStubs -> [(ForeignSrcLang, FilePath)] -> IO IfaceForeign

Convert foreign stubs and foreign files to a format suitable for writing to interfaces.

See Note [Foreign stubs and TH bytecode linking]

decodeIfaceForeign :: Logger -> TmpFs -> TempDir -> IfaceForeign -> IO (ForeignStubs, [(ForeignSrcLang, FilePath)])

Decode serialized foreign stubs and foreign files.

See Note [Foreign stubs and TH bytecode linking]