GHC.Unit.Types

Description

Unit & Module types

This module is used to resolve the loops between Unit and Module types (Module references a Unit and vice-versa).

Modules

data GenModule unit

A generic module is a pair of a unit identifier and a ModuleName.

Constructors

Module

Fields

moduleUnit :: !unit

Unit the module belongs to

moduleName :: !ModuleName

Module name (e.g. A.B.C)

type Module = GenModule Unit

A Module is a pair of a Unit and a ModuleName.

type InstalledModule = GenModule UnitId

A InstalledModule is a Module whose unit is identified with an UnitId.

type InstantiatedModule = GenModule InstantiatedUnit

An InstantiatedModule is a Module whose unit is identified with an InstantiatedUnit.

mkModule :: u -> ModuleName -> GenModule u

pprModule :: Module -> SDoc

pprInstantiatedModule :: InstantiatedModule -> SDoc

moduleFreeHoles :: GenModule (GenUnit u) -> UniqDSet ModuleName

Calculate the free holes of a Module. If this set is non-empty, this module was defined in an indefinite library that had required signatures.

If a module has free holes, that means that substitutions can operate on it; if it has no free holes, substituting over a module has no effect.

Units

data GenUnit uid

A unit identifier identifies a (possibly partially) instantiated library. It is primarily used as part of Module, which in turn is used in Name, which is used to give names to entities when typechecking.

There are two possible forms for a Unit:

1) It can be a RealUnit, in which case we just have a DefUnitId that uniquely identifies some fully compiled, installed library we have on disk.

2) It can be an VirtUnit. When we are typechecking a library with missing holes, we may need to instantiate a library on the fly (in which case we don't have any on-disk representation.) In that case, you have an InstantiatedUnit, which explicitly records the instantiation, so that we can substitute over it.

Constructors

RealUnit !(Definite uid)
Installed definite unit (either a fully instantiated unit or a closed unit)

VirtUnit !(GenInstantiatedUnit uid)
Virtual unit instantiated on-the-fly. It may be definite if all the holes are instantiated but we don't have code objects for it.

HoleUnit
Fake hole unit

type Unit = GenUnit UnitId

newtype UnitId

A UnitId identifies a built library in a database and is used to generate unique symbols, etc. It's usually of the form:

pkgname-1.2:libname+hash

These UnitId are provided to us via the -this-unit-id flag.

The library in question may be definite or indefinite; if it is indefinite, none of the holes have been filled (we never install partially instantiated libraries as we can cheaply instantiate them on-the-fly, cf VirtUnit). Put another way, an installed unit id is either fully instantiated, or not instantiated at all.

Constructors

UnitId

Fields

unitIdFS :: FastString

The full hashed unit identifier, including the component id and the hash.

data GenInstantiatedUnit unit

An instantiated unit.

It identifies an indefinite library (with holes) that has been instantiated.

This unit may be indefinite or not (i.e. with remaining holes or not). If it is definite, we don't know if it has already been compiled and installed in a database. Nevertheless, we have a mechanism called "improvement" to try to match a fully instantiated unit with existing compiled and installed units: see Note [VirtUnit to RealUnit improvement].

An indefinite unit identifier pretty-prints to something like p[H=H,A=aimpl:A>] (p is the IndefUnitId, and the brackets enclose the module substitution).

Constructors

InstantiatedUnit

Fields

instUnitFS :: !FastString

A private, uniquely identifying representation of an InstantiatedUnit. This string is completely private to GHC and is just used to get a unique.

instUnitKey :: !Unique

Cached unique of unitFS.

instUnitInstanceOf :: !(Indefinite unit)

The indefinite unit being instantiated.

instUnitInsts :: !(GenInstantiations unit)

The sorted (by ModuleName) instantiations of this unit.

instUnitHoles :: UniqDSet ModuleName

A cache of the free module holes of instUnitInsts. This lets us efficiently tell if a InstantiatedUnit has been fully instantiated (empty set of free module holes) and whether or not a substitution can have any effect.

type InstantiatedUnit = GenInstantiatedUnit UnitId

type IndefUnitId = Indefinite UnitId

An IndefUnitId is an UnitId with the invariant that it only refers to an indefinite library; i.e., one that can be instantiated.

type DefUnitId = Definite UnitId

A DefUnitId is an UnitId with the invariant that it only refers to a definite library; i.e., one we have generated code for.

type Instantiations = GenInstantiations UnitId

type GenInstantiations unit = [(ModuleName, GenModule (GenUnit unit))]

mkGenInstantiatedUnit :: (unit -> FastString) -> Indefinite unit -> GenInstantiations unit -> GenInstantiatedUnit unit

Create a new GenInstantiatedUnit given an explicit module substitution.

mkInstantiatedUnit :: IndefUnitId -> Instantiations -> InstantiatedUnit

Create a new InstantiatedUnit given an explicit module substitution.

mkInstantiatedUnitHash :: IndefUnitId -> Instantiations -> FastString

mkGenVirtUnit :: (unit -> FastString) -> Indefinite unit -> [(ModuleName, GenModule (GenUnit unit))] -> GenUnit unit

Smart constructor for instantiated GenUnit

mkVirtUnit :: IndefUnitId -> Instantiations -> Unit

Smart constructor for VirtUnit

mapGenUnit :: (u -> v) -> (v -> FastString) -> GenUnit u -> GenUnit v

Map over the unit type of a GenUnit

unitFreeModuleHoles :: GenUnit u -> UniqDSet ModuleName

Retrieve the set of free module holes of a Unit.

fsToUnit :: FastString -> Unit

Create a new simple unit identifier from a FastString. Internally, this is primarily used to specify wired-in unit identifiers.

unitFS :: Unit -> FastString

unitString :: Unit -> String

toUnitId :: Unit -> UnitId

Return the UnitId of the Unit. For on-the-fly instantiated units, return the UnitId of the indefinite unit this unit is an instance of.

virtualUnitId :: InstantiatedUnit -> UnitId

Return the virtual UnitId of an on-the-fly instantiated unit.

stringToUnit :: String -> Unit

stableUnitCmp :: Unit -> Unit -> Ordering

Compares unit ids lexically, rather than by their Uniques

unitIsDefinite :: Unit -> Bool

A Unit is definite if it has no free holes.

Unit Ids

unitIdString :: UnitId -> String

stringToUnitId :: String -> UnitId

Utils

newtype Definite unit

A definite unit (i.e. without any free module hole)

Constructors

Definite

Fields

unDefinite :: unit

data Indefinite unit

Constructors

Indefinite

Fields

indefUnit :: !unit

Unit identifier

indefUnitPprInfo :: Maybe UnitPprInfo

Cache for some unit info retrieved from the DB

Wired-in units

primUnitId :: UnitId

bignumUnitId :: UnitId

baseUnitId :: UnitId

rtsUnitId :: UnitId

thUnitId :: UnitId

mainUnitId :: UnitId

This is the package Id for the current program. It is the default package Id if you don't specify a package name. We don't add this prefix to symbol names, since there can be only one main package per program.

thisGhcUnitId :: UnitId

interactiveUnitId :: UnitId

primUnit :: Unit

bignumUnit :: Unit

baseUnit :: Unit

rtsUnit :: Unit

thUnit :: Unit

mainUnit :: Unit

thisGhcUnit :: Unit

interactiveUnit :: Unit

isInteractiveModule :: Module -> Bool

wiredInUnitIds :: [UnitId]

Boot modules

data IsBootInterface

Indicates whether a module name is referring to a boot interface (hs-boot file) or regular module (hs file). We need to treat boot modules specially when building compilation graphs, since they break cycles. Regular source files and signature files are treated equivalently.

Constructors

NotBoot

IsBoot

data GenWithIsBoot mod

This data type just pairs a value mod with an IsBootInterface flag. In practice, mod is usually a Module or ModuleName'.

Constructors

GWIB

Fields

gwib_mod :: mod

gwib_isBoot :: IsBootInterface

type ModuleNameWithIsBoot = GenWithIsBoot ModuleName

type ModuleWithIsBoot = GenWithIsBoot Module