GHC.Utils.Binary

Documentation

data Bin a

class Binary a where

Do not rely on instance sizes for general types, we use variable length encoding for many of them.

Minimal complete definition

get

Methods

put_ :: BinHandle -> a -> IO ()

put :: BinHandle -> a -> IO (Bin a)

get :: BinHandle -> IO a

data BinHandle

type SymbolTable = Array Int Name

type Dictionary = Array Int FastString

data BinData

Constructors

BinData Int BinArray

dataHandle :: BinData -> IO BinHandle

handleData :: BinHandle -> IO BinData

unsafeUnpackBinBuffer :: ByteString -> IO BinHandle

openBinMem :: Int -> IO BinHandle

seekBin :: BinHandle -> Bin a -> IO ()

tellBin :: BinHandle -> IO (Bin a)

castBin :: Bin a -> Bin b

withBinBuffer :: BinHandle -> (ByteString -> IO a) -> IO a

Get access to the underlying buffer.

foldGet :: Binary a => Word -> BinHandle -> b -> (Word -> a -> b -> IO b) -> IO b

writeBinMem :: BinHandle -> FilePath -> IO ()

readBinMem :: FilePath -> IO BinHandle

readBinMemN :: Int -> FilePath -> IO (Maybe BinHandle)

putAt :: Binary a => BinHandle -> Bin a -> a -> IO ()

getAt :: Binary a => BinHandle -> Bin a -> IO a

forwardPut :: BinHandle -> (b -> IO a) -> IO b -> IO (a, b)

"forwardPut put_A put_B" outputs A after B but allows A to be read before B by using a forward reference

forwardPut_ :: BinHandle -> (b -> IO a) -> IO b -> IO ()

forwardGet :: BinHandle -> IO a -> IO a

Read a value stored using a forward reference

For writing instances

putByte :: BinHandle -> Word8 -> IO ()

getByte :: BinHandle -> IO Word8

Variable length encodings

putULEB128 :: forall a. (Integral a, FiniteBits a) => BinHandle -> a -> IO ()

getULEB128 :: forall a. (Integral a, FiniteBits a) => BinHandle -> IO a

putSLEB128 :: forall a. (Integral a, Bits a) => BinHandle -> a -> IO ()

getSLEB128 :: forall a. (Show a, Integral a, FiniteBits a) => BinHandle -> IO a

Fixed length encoding

newtype FixedLengthEncoding a

Encode the argument in it's full length. This is different from many default binary instances which make no guarantee about the actual encoding and might do things use variable length encoding.

Constructors

FixedLengthEncoding

Fields

unFixedLength :: a

Lazy Binary I/O

lazyGet :: Binary a => BinHandle -> IO a

lazyPut :: Binary a => BinHandle -> a -> IO ()

lazyGetMaybe :: Binary a => BinHandle -> IO (Maybe a)

Deserialize a value serialized by lazyPutMaybe.

lazyPutMaybe :: Binary a => BinHandle -> Maybe a -> IO ()

Serialize the constructor strictly but lazily serialize a value inside a Just.

This way we can check for the presence of a value without deserializing the value itself.

User data

data UserData

Information we keep around during interface file serialization/deserialization. Namely we keep the functions for serializing and deserializing Names and FastStrings. We do this because we actually use serialization in two distinct settings,

When serializing interface files themselves

When computing the fingerprint of an IfaceDecl (which we computing by hashing its Binary serialization)

These two settings have different needs while serializing Names:

Names in interface files are serialized via a symbol table (see Note [Symbol table representation of names] in GHC.Iface.Binary).

During fingerprinting a binding Name is serialized as the OccName and a non-binding Name is serialized as the fingerprint of the thing they represent. See Note [Fingerprinting IfaceDecls] for further discussion.

Constructors

UserData

Fields

ud_get_name :: BinHandle -> IO Name

ud_get_fs :: BinHandle -> IO FastString

ud_put_nonbinding_name :: BinHandle -> Name -> IO ()

serialize a non-binding Name (e.g. a reference to another binding).

ud_put_binding_name :: BinHandle -> Name -> IO ()

serialize a binding Name (e.g. the name of an IfaceDecl)

ud_put_fs :: BinHandle -> FastString -> IO ()

getUserData :: BinHandle -> UserData

setUserData :: BinHandle -> UserData -> BinHandle

newReadState

Arguments

:: (BinHandle -> IO Name)
how to deserialize Names

-> (BinHandle -> IO FastString)

-> UserData

newWriteState

Arguments

:: (BinHandle -> Name -> IO ())
how to serialize non-binding Names

-> (BinHandle -> Name -> IO ())
how to serialize binding Names

-> (BinHandle -> FastString -> IO ())

-> UserData

noUserData :: UserData

String table ("dictionary")

putDictionary :: BinHandle -> Int -> UniqFM FastString (Int, FastString) -> IO ()

getDictionary :: BinHandle -> IO Dictionary

putFS :: BinHandle -> FastString -> IO ()

data FSTable

initFSTable :: BinHandle -> IO (BinHandle, FSTable, IO Int)

getDictFastString :: Dictionary -> BinHandle -> IO FastString

putDictFastString :: FSTable -> BinHandle -> FastString -> IO ()

Newtype wrappers

newtype BinSpan

Constructors

BinSpan

Fields

unBinSpan :: RealSrcSpan

newtype BinSrcSpan

Constructors

BinSrcSpan

Fields

unBinSrcSpan :: SrcSpan

newtype BinLocated a

Constructors

BinLocated

Fields

unBinLocated :: Located a