GHC.Utils.Error

Basic types

data Validity

Constructors

IsValid
Everything is fine

NotValid MsgDoc
A problem, and some indication of why

andValid :: Validity -> Validity -> Validity

allValid :: [Validity] -> Validity

If they aren't all valid, return the first

isValid :: Validity -> Bool

getInvalids :: [Validity] -> [MsgDoc]

orValid :: Validity -> Validity -> Validity

data Severity

Constructors

SevOutput

SevFatal

SevInteractive

SevDump
Log message intended for compiler developers No file/line/column stuff

SevInfo
Log messages intended for end users. No file/line/column stuff.

SevWarning

SevError
SevWarning and SevError are used for warnings and errors o The message has a file/line/column heading, plus "warning:" or "error:", added by mkLocMessags o Output is intended for end users

Messages

data ErrMsg

errMsgDoc :: ErrMsg -> ErrDoc

errMsgSeverity :: ErrMsg -> Severity

errMsgReason :: ErrMsg -> WarnReason

data ErrDoc

Categorise error msgs by their importance. This is so each section can be rendered visually distinct. See Note [Error report] for where these come from.

errDoc :: [MsgDoc] -> [MsgDoc] -> [MsgDoc] -> ErrDoc

errDocImportant :: ErrDoc -> [MsgDoc]

Primary error msg.

errDocContext :: ErrDoc -> [MsgDoc]

Context e.g. "In the second argument of ...".

errDocSupplementary :: ErrDoc -> [MsgDoc]

Supplementary information, e.g. "Relevant bindings include ...".

type WarnMsg = ErrMsg

type MsgDoc = SDoc

type Messages = (WarningMessages, ErrorMessages)

type ErrorMessages = Bag ErrMsg

type WarningMessages = Bag WarnMsg

unionMessages :: Messages -> Messages -> Messages

errMsgSpan :: ErrMsg -> SrcSpan

errMsgContext :: ErrMsg -> PrintUnqualified

errorsFound :: DynFlags -> Messages -> Bool

isEmptyMessages :: Messages -> Bool

isWarnMsgFatal :: DynFlags -> WarnMsg -> Maybe (Maybe WarningFlag)

Checks if given WarnMsg is a fatal warning.

warningsToMessages :: DynFlags -> WarningMessages -> Messages

Formatting

pprMessageBag :: Bag MsgDoc -> SDoc

pprErrMsgBagWithLoc :: Bag ErrMsg -> [SDoc]

pprLocErrMsg :: ErrMsg -> SDoc

printBagOfErrors :: DynFlags -> Bag ErrMsg -> IO ()

formatErrDoc :: SDocContext -> ErrDoc -> SDoc

Construction

emptyMessages :: Messages

mkLocMessage :: Severity -> SrcSpan -> MsgDoc -> MsgDoc

Make an unannotated error message with location info.

mkLocMessageAnn

Arguments

:: Maybe String
optional annotation

-> Severity
severity

-> SrcSpan
location

-> MsgDoc
message

-> MsgDoc

Make a possibly annotated error message with location info.

makeIntoWarning :: WarnReason -> ErrMsg -> ErrMsg

mkErrMsg :: DynFlags -> SrcSpan -> PrintUnqualified -> MsgDoc -> ErrMsg

A short (one-line) error message

mkPlainErrMsg :: DynFlags -> SrcSpan -> MsgDoc -> ErrMsg

Variant that doesn't care about qualified/unqualified names

mkErrDoc :: DynFlags -> SrcSpan -> PrintUnqualified -> ErrDoc -> ErrMsg

mkLongErrMsg :: DynFlags -> SrcSpan -> PrintUnqualified -> MsgDoc -> MsgDoc -> ErrMsg

A long (multi-line) error message

mkWarnMsg :: DynFlags -> SrcSpan -> PrintUnqualified -> MsgDoc -> ErrMsg

A short (one-line) error message

mkPlainWarnMsg :: DynFlags -> SrcSpan -> MsgDoc -> ErrMsg

Variant that doesn't care about qualified/unqualified names

mkLongWarnMsg :: DynFlags -> SrcSpan -> PrintUnqualified -> MsgDoc -> MsgDoc -> ErrMsg

A long (multi-line) error message

Utilities

doIfSet :: Bool -> IO () -> IO ()

doIfSet_dyn :: DynFlags -> GeneralFlag -> IO () -> IO ()

getCaretDiagnostic :: Severity -> SrcSpan -> IO MsgDoc

Dump files

dumpIfSet :: DynFlags -> Bool -> String -> SDoc -> IO ()

dumpIfSet_dyn :: DynFlags -> DumpFlag -> String -> DumpFormat -> SDoc -> IO ()

A wrapper around dumpAction. First check whether the dump flag is set Do nothing if it is unset

dumpIfSet_dyn_printer :: PrintUnqualified -> DynFlags -> DumpFlag -> String -> DumpFormat -> SDoc -> IO ()

A wrapper around dumpAction. First check whether the dump flag is set Do nothing if it is unset

Unlike dumpIfSet_dyn, has a printer argument

dumpOptionsFromFlag :: DumpFlag -> DumpOptions

Create dump options from a DumpFlag

data DumpOptions

Dump options

Dumps are printed on stdout by default except when the dumpForcedToFile field is set to True.

When dumpForcedToFile is True or when `-ddump-to-file` is set, dumps are written into a file whose suffix is given in the dumpSuffix field.

Constructors

DumpOptions

Fields

dumpForcedToFile :: Bool

Must be dumped into a file, even if -ddump-to-file isn't set

dumpSuffix :: String

Filename suffix used when dumped into a file

data DumpFormat

Format of a dump

Dump formats are loosely defined: dumps may contain various additional headers and annotations and they may be partial. DumpFormat is mainly a hint (e.g. for syntax highlighters).

Constructors

FormatHaskell
Haskell

FormatCore
Core

FormatSTG
STG

FormatByteCode
ByteCode

FormatCMM
Cmm

FormatASM
Assembly code

FormatC
C code/header

FormatLLVM
LLVM bytecode

FormatText
Unstructured dump

type DumpAction = DynFlags -> PprStyle -> DumpOptions -> String -> DumpFormat -> SDoc -> IO ()

dumpAction :: DumpAction

Helper for dump_action

defaultDumpAction :: DumpAction

Default action for dumpAction hook

type TraceAction = forall a. DynFlags -> String -> SDoc -> a -> a

traceAction :: TraceAction

Helper for trace_action

defaultTraceAction :: TraceAction

Default action for traceAction hook

touchDumpFile :: DynFlags -> DumpOptions -> IO ()

Ensure that a dump file is created even if it stays empty

Issuing messages during compilation

putMsg :: DynFlags -> MsgDoc -> IO ()

printInfoForUser :: DynFlags -> PrintUnqualified -> MsgDoc -> IO ()

printOutputForUser :: DynFlags -> PrintUnqualified -> MsgDoc -> IO ()

logInfo :: DynFlags -> MsgDoc -> IO ()

logOutput :: DynFlags -> MsgDoc -> IO ()

Like logInfo but with SevOutput rather then SevInfo

errorMsg :: DynFlags -> MsgDoc -> IO ()

warningMsg :: DynFlags -> MsgDoc -> IO ()

fatalErrorMsg :: DynFlags -> MsgDoc -> IO ()

fatalErrorMsg'' :: FatalMessager -> String -> IO ()

compilationProgressMsg :: DynFlags -> String -> IO ()

showPass :: DynFlags -> String -> IO ()

withTiming

Arguments

:: MonadIO m

=> DynFlags
DynFlags

-> SDoc
The name of the phase

-> (a -> ())
A function to force the result (often either const () or rnf)

-> m a
The body of the phase to be timed

-> m a

Time a compilation phase.

When timings are enabled (e.g. with the -v2 flag), the allocations and CPU time used by the phase will be reported to stderr. Consider a typical usage: withTiming getDynFlags (text "simplify") force PrintTimings pass. When timings are enabled the following costs are included in the produced accounting,

The cost of executing pass to a result r in WHNF

The cost of evaluating force r to WHNF (e.g. ())

The choice of the force function depends upon the amount of forcing desired; the goal here is to ensure that the cost of evaluating the result is, to the greatest extent possible, included in the accounting provided by withTiming. Often the pass already sufficiently forces its result during construction; in this case const () is a reasonable choice. In other cases, it is necessary to evaluate the result to normal form, in which case something like Control.DeepSeq.rnf is appropriate.

To avoid adversely affecting compiler performance when timings are not requested, the result is only forced when timings are enabled.

See Note [withTiming] for more.

withTimingSilent

Arguments

:: MonadIO m

=> DynFlags
DynFlags

-> SDoc
The name of the phase

-> (a -> ())
A function to force the result (often either const () or rnf)

-> m a
The body of the phase to be timed

-> m a

Same as withTiming, but doesn't print timings in the console (when given -vN, N >= 2 or -ddump-timings).

See Note [withTiming] for more.

withTimingD

Arguments

:: (MonadIO m, HasDynFlags m)

=> SDoc
The name of the phase

-> (a -> ())
A function to force the result (often either const () or rnf)

-> m a
The body of the phase to be timed

-> m a

Like withTiming but get DynFlags from the Monad.

withTimingSilentD

Arguments

:: (MonadIO m, HasDynFlags m)

=> SDoc
The name of the phase

-> (a -> ())
A function to force the result (often either const () or rnf)

-> m a
The body of the phase to be timed

-> m a

Same as withTiming, but doesn't print timings in the console (when given -vN, N >= 2 or -ddump-timings) and gets the DynFlags from the given Monad.

See Note [withTiming] for more.

debugTraceMsg :: DynFlags -> Int -> MsgDoc -> IO ()

ghcExit :: DynFlags -> Int -> IO ()

prettyPrintGhcErrors :: ExceptionMonad m => DynFlags -> m a -> m a

traceCmd :: DynFlags -> String -> String -> IO a -> IO a