GHC.Utils.Monad.State.Strict

Description

A state monad which is strict in its state.

The State monad

data State s a where

A state monad which is strict in the state s, but lazy in the value a.

See Note [Strict State monad] for the particular notion of strictness and implementation details.

Bundled Patterns

pattern State :: (s -> (# a, s #)) -> State s a

state :: (s -> (a, s)) -> State s a

evalState :: State s a -> s -> a

execState :: State s a -> s -> s

runState :: State s a -> s -> (a, s)

Operations

get :: State s s

gets :: (s -> a) -> State s a

put :: s -> State s ()

modify :: (s -> s) -> State s ()