GHC.Utils.Outputable

Description

This module defines classes and functions for pretty-printing. It also exports a number of helpful debugging and other utilities such as trace and panic.

The interface to this module is very similar to the standard Hughes-PJ pretty printing module, except that it exports a number of additional functions that are rarely used, and works over the SDoc type.

Type classes

class Outputable a where

Class designating that some type has an SDoc representation

Minimal complete definition

Nothing

Methods

ppr :: a -> SDoc

pprPrec :: Rational -> a -> SDoc

class Outputable a => OutputableBndr a where

When we print a binder, we often want to print its type too. The OutputableBndr class encapsulates this idea.

Minimal complete definition

pprPrefixOcc, pprInfixOcc

Methods

pprBndr :: BindingSite -> a -> SDoc

pprPrefixOcc :: a -> SDoc

pprInfixOcc :: a -> SDoc

bndrIsJoin_maybe :: a -> Maybe Int

Pretty printing combinators

data SDoc

Represents a pretty-printable document.

To display an SDoc, use printSDoc, printSDocLn, bufLeftRenderSDoc, or renderWithStyle. Avoid calling runSDoc directly as it breaks the abstraction layer.

runSDoc :: SDoc -> SDocContext -> Doc

initSDocContext :: DynFlags -> PprStyle -> SDocContext

docToSDoc :: Doc -> SDoc

interppSP :: Outputable a => [a] -> SDoc

Returns the separated concatenation of the pretty printed things.

interpp'SP :: Outputable a => [a] -> SDoc

Returns the comma-separated concatenation of the pretty printed things.

pprQuotedList :: Outputable a => [a] -> SDoc

Returns the comma-separated concatenation of the quoted pretty printed things.

[x,y,z]  ==>  `x', `y', `z'

pprWithCommas

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> [a]
The things to be pretty printed

-> SDoc
SDoc where the things have been pretty printed, comma-separated and finally packed into a paragraph.

quotedListWithOr :: [SDoc] -> SDoc

quotedListWithNor :: [SDoc] -> SDoc

pprWithBars

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> [a]
The things to be pretty printed

-> SDoc
SDoc where the things have been pretty printed, bar-separated and finally packed into a paragraph.

empty :: SDoc

isEmpty :: SDocContext -> SDoc -> Bool

nest :: Int -> SDoc -> SDoc

Indent SDoc some specified amount

char :: Char -> SDoc

text :: String -> SDoc

ftext :: FastString -> SDoc

ptext :: PtrString -> SDoc

ztext :: FastZString -> SDoc

int :: Int -> SDoc

intWithCommas :: Integral a => a -> SDoc

integer :: Integer -> SDoc

word :: Integer -> SDoc

float :: Float -> SDoc

double :: Double -> SDoc

rational :: Rational -> SDoc

doublePrec :: Int -> Double -> SDoc

doublePrec p n shows a floating point number n with p digits of precision after the decimal point.

parens :: SDoc -> SDoc

cparen :: Bool -> SDoc -> SDoc

brackets :: SDoc -> SDoc

braces :: SDoc -> SDoc

quotes :: SDoc -> SDoc

quote :: SDoc -> SDoc

doubleQuotes :: SDoc -> SDoc

angleBrackets :: SDoc -> SDoc

semi :: SDoc

comma :: SDoc

colon :: SDoc

dcolon :: SDoc

space :: SDoc

equals :: SDoc

dot :: SDoc

vbar :: SDoc

arrow :: SDoc

lollipop :: SDoc

larrow :: SDoc

darrow :: SDoc

arrowt :: SDoc

larrowt :: SDoc

arrowtt :: SDoc

larrowtt :: SDoc

lparen :: SDoc

rparen :: SDoc

lbrack :: SDoc

rbrack :: SDoc

lbrace :: SDoc

rbrace :: SDoc

underscore :: SDoc

mulArrow :: SDoc -> SDoc

blankLine :: SDoc

forAllLit :: SDoc

bullet :: SDoc

(<>) :: SDoc -> SDoc -> SDoc

Join two SDoc together horizontally without a gap

(<+>) :: SDoc -> SDoc -> SDoc

Join two SDoc together horizontally with a gap between them

hcat :: [SDoc] -> SDoc

Concatenate SDoc horizontally

hsep :: [SDoc] -> SDoc

Concatenate SDoc horizontally with a space between each one

($$) :: SDoc -> SDoc -> SDoc

Join two SDoc together vertically; if there is no vertical overlap it "dovetails" the two onto one line

($+$) :: SDoc -> SDoc -> SDoc

Join two SDoc together vertically

vcat :: [SDoc] -> SDoc

Concatenate SDoc vertically with dovetailing

sep :: [SDoc] -> SDoc

Separate: is either like hsep or like vcat, depending on what fits

cat :: [SDoc] -> SDoc

Catenate: is either like hcat or like vcat, depending on what fits

fsep :: [SDoc] -> SDoc

A paragraph-fill combinator. It's much like sep, only it keeps fitting things on one line until it can't fit any more.

fcat :: [SDoc] -> SDoc

This behaves like fsep, but it uses <> for horizontal conposition rather than <+>

hang

Arguments

:: SDoc
The header

-> Int
Amount to indent the hung body

-> SDoc
The hung body, indented and placed below the header

-> SDoc

hangNotEmpty :: SDoc -> Int -> SDoc -> SDoc

This behaves like hang, but does not indent the second document when the header is empty.

punctuate

Arguments

:: SDoc
The punctuation

-> [SDoc]
The list that will have punctuation added between every adjacent pair of elements

-> [SDoc]
Punctuated list

ppWhen :: Bool -> SDoc -> SDoc

ppUnless :: Bool -> SDoc -> SDoc

ppWhenOption :: (SDocContext -> Bool) -> SDoc -> SDoc

ppUnlessOption :: (SDocContext -> Bool) -> SDoc -> SDoc

speakNth :: Int -> SDoc

Converts an integer to a verbal index:

speakNth 1 = text "first"
speakNth 5 = text "fifth"
speakNth 21 = text "21st"

speakN :: Int -> SDoc

Converts an integer to a verbal multiplicity:

speakN 0 = text "none"
speakN 5 = text "five"
speakN 10 = text "10"

speakNOf :: Int -> SDoc -> SDoc

Converts an integer and object description to a statement about the multiplicity of those objects:

speakNOf 0 (text "melon") = text "no melons"
speakNOf 1 (text "melon") = text "one melon"
speakNOf 3 (text "melon") = text "three melons"

plural :: [a] -> SDoc

Determines the pluralisation suffix appropriate for the length of a list:

plural [] = char 's'
plural ["Hello"] = empty
plural ["Hello", "World"] = char 's'

isOrAre :: [a] -> SDoc

Determines the form of to be appropriate for the length of a list:

isOrAre [] = text "are"
isOrAre ["Hello"] = text "is"
isOrAre ["Hello", "World"] = text "are"

doOrDoes :: [a] -> SDoc

Determines the form of to do appropriate for the length of a list:

doOrDoes [] = text "do"
doOrDoes ["Hello"] = text "does"
doOrDoes ["Hello", "World"] = text "do"

itsOrTheir :: [a] -> SDoc

Determines the form of possessive appropriate for the length of a list:

itsOrTheir [x]   = text "its"
itsOrTheir [x,y] = text "their"
itsOrTheir []    = text "their"  -- probably avoid this

unicodeSyntax :: SDoc -> SDoc -> SDoc

coloured :: PprColour -> SDoc -> SDoc

Apply the given colour/style for the argument.

Only takes effect if colours are enabled.

keyword :: SDoc -> SDoc

Converting SDoc into strings and outputting it

printSDoc :: SDocContext -> Mode -> Handle -> SDoc -> IO ()

The analog of printDoc_ for SDoc, which tries to make sure the terminal doesn't get screwed up by the ANSI color codes if an exception is thrown during pretty-printing.

printSDocLn :: SDocContext -> Mode -> Handle -> SDoc -> IO ()

Like printSDoc but appends an extra newline.

printForUser :: DynFlags -> Handle -> PrintUnqualified -> Depth -> SDoc -> IO ()

printForC :: DynFlags -> Handle -> SDoc -> IO ()

Like printSDocLn but specialized with LeftMode and PprCode CStyle. This is typically used to output C-- code.

bufLeftRenderSDoc :: SDocContext -> BufHandle -> SDoc -> IO ()

An efficient variant of printSDoc specialized for LeftMode that outputs to a BufHandle.

pprCode :: CodeStyle -> SDoc -> SDoc

mkCodeStyle :: CodeStyle -> PprStyle

showSDoc :: DynFlags -> SDoc -> String

showSDocUnsafe :: SDoc -> String

showSDocOneLine :: SDocContext -> SDoc -> String

showSDocForUser :: DynFlags -> PrintUnqualified -> SDoc -> String

showSDocDebug :: DynFlags -> SDoc -> String

showSDocDump :: DynFlags -> SDoc -> String

showSDocDumpOneLine :: DynFlags -> SDoc -> String

showSDocUnqual :: DynFlags -> SDoc -> String

showPpr :: Outputable a => DynFlags -> a -> String

renderWithStyle :: SDocContext -> SDoc -> String

pprInfixVar :: Bool -> SDoc -> SDoc

pprPrefixVar :: Bool -> SDoc -> SDoc

pprHsChar :: Char -> SDoc

Special combinator for showing character literals.

pprHsString :: FastString -> SDoc

Special combinator for showing string literals.

pprHsBytes :: ByteString -> SDoc

Special combinator for showing bytestring literals.

primFloatSuffix :: SDoc

primCharSuffix :: SDoc

primWordSuffix :: SDoc

primDoubleSuffix :: SDoc

primInt64Suffix :: SDoc

primWord64Suffix :: SDoc

primIntSuffix :: SDoc

pprPrimChar :: Char -> SDoc

Special combinator for showing unboxed literals.

pprPrimInt :: Integer -> SDoc

pprPrimWord :: Integer -> SDoc

pprPrimInt64 :: Integer -> SDoc

pprPrimWord64 :: Integer -> SDoc

pprFastFilePath :: FastString -> SDoc

pprFilePathString :: FilePath -> SDoc

Normalise, escape and render a string representing a path

e.g. "c:\whatever"

Controlling the style in which output is printed

data BindingSite

BindingSite is used to tell the thing that prints binder what language construct is binding the identifier. This can be used to decide how much info to print. Also see Note [Binding-site specific printing] in GHC.Core.Ppr

Constructors

LambdaBind
The x in (x. e)

CaseBind
The x in case scrut of x { (y,z) -> ... }

CasePatBind
The y,z in case scrut of x { (y,z) -> ... }

LetBind
The x in (let x = rhs in e)

data PprStyle

Constructors

PprUser PrintUnqualified Depth Coloured

PprDump PrintUnqualified

PprCode CodeStyle

data CodeStyle

Constructors

CStyle

AsmStyle

data PrintUnqualified

When printing code that contains original names, we need to map the original names back to something the user understands. This is the purpose of the triple of functions that gets passed around when rendering SDoc.

Constructors

QueryQualify

Fields

queryQualifyName :: QueryQualifyName

queryQualifyModule :: QueryQualifyModule

queryQualifyPackage :: QueryQualifyPackage

type QueryQualifyName = Module -> OccName -> QualifyName

Given a Name's Module and OccName, decide whether and how to qualify it.

type QueryQualifyModule = Module -> Bool

For a given module, we need to know whether to print it with a package name to disambiguate it.

type QueryQualifyPackage = Unit -> Bool

For a given package, we need to know whether to print it with the component id to disambiguate it.

reallyAlwaysQualify :: PrintUnqualified

reallyAlwaysQualifyNames :: QueryQualifyName

alwaysQualify :: PrintUnqualified

alwaysQualifyNames :: QueryQualifyName

NB: This won't ever show package IDs

alwaysQualifyModules :: QueryQualifyModule

neverQualify :: PrintUnqualified

neverQualifyNames :: QueryQualifyName

neverQualifyModules :: QueryQualifyModule

alwaysQualifyPackages :: QueryQualifyPackage

neverQualifyPackages :: QueryQualifyPackage

data QualifyName

Constructors

NameUnqual

NameQual ModuleName

NameNotInScope1

NameNotInScope2

queryQual :: PprStyle -> PrintUnqualified

sdocWithDynFlags :: (DynFlags -> SDoc) -> SDoc

sdocOption :: (SDocContext -> a) -> (a -> SDoc) -> SDoc

updSDocContext :: (SDocContext -> SDocContext) -> SDoc -> SDoc

data SDocContext

Constructors

SDC

Fields

sdocStyle :: !PprStyle

sdocColScheme :: !Scheme

sdocLastColour :: !PprColour

The most recently used colour. This allows nesting colours.

sdocShouldUseColor :: !Bool

sdocDefaultDepth :: !Int

sdocLineLength :: !Int

sdocCanUseUnicode :: !Bool

True if Unicode encoding is supported and not disable by GHC_NO_UNICODE environment variable

sdocHexWordLiterals :: !Bool

sdocPprDebug :: !Bool

sdocPrintUnicodeSyntax :: !Bool

sdocPrintCaseAsLet :: !Bool

sdocPrintTypecheckerElaboration :: !Bool

sdocPrintAxiomIncomps :: !Bool

sdocPrintExplicitKinds :: !Bool

sdocPrintExplicitCoercions :: !Bool

sdocPrintExplicitRuntimeReps :: !Bool

sdocPrintExplicitForalls :: !Bool

sdocPrintPotentialInstances :: !Bool

sdocPrintEqualityRelations :: !Bool

sdocSuppressTicks :: !Bool

sdocSuppressTypeSignatures :: !Bool

sdocSuppressTypeApplications :: !Bool

sdocSuppressIdInfo :: !Bool

sdocSuppressCoercions :: !Bool

sdocSuppressUnfoldings :: !Bool

sdocSuppressVarKinds :: !Bool

sdocSuppressUniques :: !Bool

sdocSuppressModulePrefixes :: !Bool

sdocSuppressStgExts :: !Bool

sdocErrorSpans :: !Bool

sdocStarIsType :: !Bool

sdocLinearTypes :: !Bool

sdocImpredicativeTypes :: !Bool

sdocPrintTypeAbbreviations :: !Bool

sdocDynFlags :: DynFlags

sdocWithContext :: (SDocContext -> SDoc) -> SDoc

getPprStyle :: (PprStyle -> SDoc) -> SDoc

withPprStyle :: PprStyle -> SDoc -> SDoc

setStyleColoured :: Bool -> PprStyle -> PprStyle

pprDeeper :: SDoc -> SDoc

pprDeeperList :: ([SDoc] -> SDoc) -> [SDoc] -> SDoc

Truncate a list that is longer than the current depth.

pprSetDepth :: Depth -> SDoc -> SDoc

codeStyle :: PprStyle -> Bool

userStyle :: PprStyle -> Bool

dumpStyle :: PprStyle -> Bool

asmStyle :: PprStyle -> Bool

qualName :: PprStyle -> QueryQualifyName

qualModule :: PprStyle -> QueryQualifyModule

qualPackage :: PprStyle -> QueryQualifyPackage

mkErrStyle :: PrintUnqualified -> PprStyle

Style for printing error messages

defaultErrStyle :: PprStyle

Default style for error messages, when we don't know PrintUnqualified It's a bit of a hack because it doesn't take into account what's in scope Only used for desugarer warnings, and typechecker errors in interface sigs

defaultDumpStyle :: PprStyle

mkDumpStyle :: PrintUnqualified -> PprStyle

defaultUserStyle :: PprStyle

mkUserStyle :: PrintUnqualified -> Depth -> PprStyle

cmdlineParserStyle :: PprStyle

data Depth

Constructors

AllTheWay

PartWay Int
0 => stop

DefaultDepth
Use sdocDefaultDepth field as depth

withUserStyle :: PrintUnqualified -> Depth -> SDoc -> SDoc

withErrStyle :: PrintUnqualified -> SDoc -> SDoc

ifPprDebug :: SDoc -> SDoc -> SDoc

Says what to do with and without -dppr-debug

whenPprDebug :: SDoc -> SDoc

Says what to do with -dppr-debug; without, return empty

getPprDebug :: (Bool -> SDoc) -> SDoc

Indicate if -dppr-debug mode is enabled

Error handling and debugging utilities

pprPanic :: HasCallStack => String -> SDoc -> a

Throw an exception saying "bug in GHC"

pprSorry :: String -> SDoc -> a

Throw an exception saying "this isn't finished yet"

assertPprPanic :: HasCallStack => String -> Int -> SDoc -> a

Panic with an assertion failure, recording the given file and line number. Should typically be accessed with the ASSERT family of macros

pprPgmError :: String -> SDoc -> a

Throw an exception saying "bug in pgm being compiled" (used for unusual program errors)

pprTrace :: String -> SDoc -> a -> a

If debug output is on, show some SDoc on the screen

pprTraceDebug :: String -> SDoc -> a -> a

pprTraceWith :: String -> (a -> SDoc) -> a -> a

pprTraceWith desc f x is equivalent to pprTrace desc (f x) x. This allows you to print details from the returned value as well as from ambient variables.

pprTraceIt :: Outputable a => String -> a -> a

pprTraceIt desc x is equivalent to pprTrace desc (ppr x) x

warnPprTrace :: HasCallStack => Bool -> String -> Int -> SDoc -> a -> a

Just warn about an assertion failure, recording the given file and line number. Should typically be accessed with the WARN macros

pprSTrace :: HasCallStack => SDoc -> a -> a

If debug output is on, show some SDoc on the screen along with a call stack when available.

pprTraceException :: ExceptionMonad m => String -> SDoc -> m a -> m a

pprTraceException desc x action runs action, printing a message if it throws an exception.

pprTraceM :: Applicative f => String -> SDoc -> f ()

pprTraceWithFlags :: DynFlags -> String -> SDoc -> a -> a

If debug output is on, show some SDoc on the screen

trace :: String -> a -> a

The trace function outputs the trace message given as its first argument, before returning the second argument as its result.

For example, this returns the value of f x but first outputs the message.

>>> let x = 123; f = show
>>> trace ("calling f with x = " ++ show x) (f x)
"calling f with x = 123
123"

The trace function should only be used for debugging, or for monitoring execution. The function is not referentially transparent: its type indicates that it is a pure function but it has the side effect of outputting the trace message.

pgmError :: String -> a

Panics and asserts.

panic :: String -> a

Panics and asserts.

sorry :: String -> a

Panics and asserts.

assertPanic :: String -> Int -> a

Throw a failed assertion exception for a given filename and line number.

pprDebugAndThen :: DynFlags -> (String -> a) -> SDoc -> SDoc -> a

callStackDoc :: HasCallStack => SDoc