GhcPlugins

Description

This module is not used by GHC itself. Rather, it exports all of the functions and types you are likely to need when writing a plugin for GHC. So authors of plugins can probably get away simply with saying "import GhcPlugins".

Particularly interesting modules for plugin writers include CoreSyn and CoreMonad.

Documentation

module Plugins

module RdrName

module OccName

module Name

module Var

module Id

module IdInfo

module CoreMonad

module CoreSyn

module Literal

module DataCon

module CoreUtils

module MkCore

module CoreFVs

module CoreSubst

module Rules

module Annotations

module DynFlags

module Packages

module Module

module Type

module TyCon

module Coercion

module TysWiredIn

module HscTypes

module BasicTypes

module VarSet

module VarEnv

module NameSet

module NameEnv

module UniqSet

module UniqFM

module FiniteMap

module Util

module GHC.Serialized

module SrcLoc

module Outputable

module UniqSupply

module Unique

module FastString