GraphPpr

Description

Pretty printing of graphs.

Documentation

dumpGraph :: (Outputable k, Outputable color) => Graph k cls color -> SDoc

Pretty print a graph in a somewhat human readable format.

dotGraph

Arguments

:: (Uniquable k, Outputable k, Outputable cls, Outputable color)

=> (color -> SDoc)
What graphviz color to use for each node color It's usually safe to return X11 style colors here, ie "red", "green" etc or a hex triplet #aaff55 etc

-> Triv k cls color

-> Graph k cls color

-> SDoc

Pretty print a graph in graphviz .dot format. Conflicts get solid edges. Coalescences get dashed edges.