Hoopl.Block

Documentation

data C

Used at the type level to indicate a "closed" structure which supports control transfer only through the use of named labels---no "fallthrough" is permitted. The number of control-flow edges is unconstrained.

data O

Used at the type level to indicate an "open" structure with a unique, unnamed control-flow edge flowing in or out. Fallthrough and concatenation are permitted at an open point.

data MaybeO ex t where

Maybe type indexed by open/closed

Constructors

JustO :: t -> MaybeO O t

NothingO :: MaybeO C t

type family IndexedCO ex a b :: *

Either type indexed by closed/open using type families

data Block n e x where

A sequence of nodes. May be any of four shapes (OO, OC, CO, CC). Open at the entry means single entry, mutatis mutandis for exit. A closedclosed block is a basic/ block and can't be extended further. Clients should avoid manipulating blocks and should stick to either nodes or graphs.

Constructors

BlockCO :: n C O -> Block n O O -> Block n C O

BlockCC :: n C O -> Block n O O -> n O C -> Block n C C

BlockOC :: Block n O O -> n O C -> Block n O C

BNil :: Block n O O

BMiddle :: n O O -> Block n O O

BCat :: Block n O O -> Block n O O -> Block n O O

BSnoc :: Block n O O -> n O O -> Block n O O

BCons :: n O O -> Block n O O -> Block n O O

blockAppend :: Block n e O -> Block n O x -> Block n e x

blockCons :: n O O -> Block n O x -> Block n O x

blockFromList :: [n O O] -> Block n O O

blockJoin :: n C O -> Block n O O -> n O C -> Block n C C

blockJoinHead :: n C O -> Block n O x -> Block n C x

blockJoinTail :: Block n e O -> n O C -> Block n e C

blockSnoc :: Block n e O -> n O O -> Block n e O

blockSplit :: Block n C C -> (n C O, Block n O O, n O C)

Split a closed block into its entry node, open middle block, and exit node.

blockSplitHead :: Block n C x -> (n C O, Block n O x)

blockSplitTail :: Block n e C -> (Block n e O, n O C)

blockToList :: Block n O O -> [n O O]

emptyBlock :: Block n O O

firstNode :: Block n C x -> n C O

foldBlockNodesB :: forall n a. (forall e x. n e x -> a -> a) -> forall e x. Block n e x -> IndexedCO x a a -> IndexedCO e a a

foldBlockNodesB3 :: forall n a b c. (n C O -> b -> c, n O O -> b -> b, n O C -> a -> b) -> forall e x. Block n e x -> IndexedCO x a b -> IndexedCO e c b

foldBlockNodesF :: forall n a. (forall e x. n e x -> a -> a) -> forall e x. Block n e x -> IndexedCO e a a -> IndexedCO x a a

isEmptyBlock :: Block n e x -> Bool

lastNode :: Block n x C -> n O C

mapBlock :: (forall e x. n e x -> n' e x) -> Block n e x -> Block n' e x

map a function over the nodes of a Block

mapBlock' :: (forall e x. n e x -> n' e x) -> Block n e x -> Block n' e x

A strict mapBlock

mapBlock3' :: forall n n' e x. (n C O -> n' C O, n O O -> n' O O, n O C -> n' O C) -> Block n e x -> Block n' e x

map over a block, with different functions to apply to first nodes, middle nodes and last nodes respectively. The map is strict.

replaceFirstNode :: Block n C x -> n C O -> Block n C x

replaceLastNode :: Block n x C -> n O C -> Block n x C