HsBinds

Documentation

type HsLocalBinds id = HsLocalBindsLR id id

Haskell Local Bindings

type LHsLocalBinds id = Located (HsLocalBinds id)

Located Haskell local bindings

data HsLocalBindsLR idL idR

Haskell Local Bindings with separate Left and Right identifier types

Bindings in a 'let' expression or a 'where' clause

Constructors

HsValBinds (XHsValBinds idL idR) (HsValBindsLR idL idR)
Haskell Value Bindings

HsIPBinds (XHsIPBinds idL idR) (HsIPBinds idR)
Haskell Implicit Parameter Bindings

EmptyLocalBinds (XEmptyLocalBinds idL idR)
Empty Local Bindings

XHsLocalBindsLR (XXHsLocalBindsLR idL idR)

type LHsLocalBindsLR idL idR = Located (HsLocalBindsLR idL idR)

type HsValBinds id = HsValBindsLR id id

Haskell Value Bindings

data HsValBindsLR idL idR

Haskell Value bindings with separate Left and Right identifier types (not implicit parameters) Used for both top level and nested bindings May contain pattern synonym bindings

Constructors

ValBinds (XValBinds idL idR) (LHsBindsLR idL idR) [LSig idR]
Value Bindings In

Before renaming RHS; idR is always RdrName Not dependency analysed Recursive by default

XValBindsLR (XXValBindsLR idL idR)
Value Bindings Out

After renaming RHS; idR can be Name or Id Dependency analysed, later bindings in the list may depend on earlier ones.

data NHsValBindsLR idL

Constructors

NValBinds [(RecFlag, LHsBinds idL)] [LSig GhcRn]

type LHsBind id = LHsBindLR id id

Located Haskell Binding

type LHsBinds id = LHsBindsLR id id

Located Haskell Bindings

type HsBind id = HsBindLR id id

Haskell Binding

type LHsBindsLR idL idR = Bag (LHsBindLR idL idR)

Located Haskell Bindings with separate Left and Right identifier types

type LHsBindLR idL idR = Located (HsBindLR idL idR)

Located Haskell Binding with separate Left and Right identifier types

data HsBindLR idL idR

Haskell Binding with separate Left and Right id's

Constructors

FunBind
Function-like Binding

FunBind is used for both functions f x = e and variables f = x -> e and strict variables !x = x + 1

Reason 1: Special case for type inference: see tcMonoBinds.

Reason 2: Instance decls can only have FunBinds, which is convenient. If you change this, you'll need to change e.g. rnMethodBinds

But note that the form f :: a->a = ... parses as a pattern binding, just like (f :: a -> a) = ...

Strict bindings have their strictness recorded in the SrcStrictness of their MatchContext. See Note [FunBind vs PatBind] for details about the relationship between FunBind and PatBind.

AnnKeywordIds

AnnFunId, attached to each element of fun_matches

AnnEqual,AnnWhere, AnnOpen,AnnClose,

Fields

fun_ext :: XFunBind idL idR

After the renamer, this contains the locally-bound free variables of this defn. See Note [Bind free vars]

fun_id :: Located (IdP idL)

fun_matches :: MatchGroup idR (LHsExpr idR)

The payload

fun_co_fn :: HsWrapper

Coercion from the type of the MatchGroup to the type of the Id. Example:

     f :: Int -> forall a. a -> a
     f x y = y

Then the MatchGroup will have type (Int -> a' -> a') (with a free type variable a'). The coercion will take a CoreExpr of this type and convert it to a CoreExpr of type Int -> forall a'. a' -> a' Notice that the coercion captures the free a'.

fun_tick :: [Tickish Id]

Ticks to put on the rhs, if any

PatBind
Pattern Binding

The pattern is never a simple variable; That case is done by FunBind. See Note [FunBind vs PatBind] for details about the relationship between FunBind and PatBind.

Fields

pat_ext :: XPatBind idL idR

See Note [Bind free vars]

pat_lhs :: LPat idL

pat_rhs :: GRHSs idR (LHsExpr idR)

pat_ticks :: ([Tickish Id], [[Tickish Id]])

Ticks to put on the rhs, if any, and ticks to put on the bound variables.

VarBind
Variable Binding

Dictionary binding and suchlike. All VarBinds are introduced by the type checker

Fields

var_ext :: XVarBind idL idR

var_id :: IdP idL

var_rhs :: LHsExpr idR

Located only for consistency

var_inline :: Bool

True = inline this binding regardless (used for implication constraints only)

AbsBinds
Abstraction Bindings

Fields

abs_ext :: XAbsBinds idL idR

abs_tvs :: [TyVar]

abs_ev_vars :: [EvVar]

Includes equality constraints

abs_exports :: [ABExport idL]

AbsBinds only gets used when idL = idR after renaming, but these need to be idL's for the collect... code in HsUtil to have the right type

abs_ev_binds :: [TcEvBinds]

Evidence bindings Why a list? See TcInstDcls Note [Typechecking plan for instance declarations]

abs_binds :: LHsBinds idL

Typechecked user bindings

abs_sig :: Bool

PatSynBind (XPatSynBind idL idR) (PatSynBind idL idR)

AnnKeywordId : AnnPattern, AnnLarrow,AnnEqual, AnnWhere AnnOpen '{',AnnClose '}'

XHsBindsLR (XXHsBindsLR idL idR)

data NPatBindTc

Constructors

NPatBindTc

Fields

pat_fvs :: NameSet

Free variables

pat_rhs_ty :: Type

Type of the GRHSs

data ABExport p

Abtraction Bindings Export

Constructors

ABE

Fields

abe_ext :: XABE p

abe_poly :: IdP p

Any INLINE pragma is attached to this Id

abe_mono :: IdP p

abe_wrap :: HsWrapper

See Note [ABExport wrapper] Shape: (forall abs_tvs. abs_ev_vars => abe_mono) ~ abe_poly

abe_prags :: TcSpecPrags

SPECIALISE pragmas

XABExport (XXABExport p)

data PatSynBind idL idR

AnnKeywordId : AnnPattern, AnnEqual,AnnLarrow AnnWhere,AnnOpen '{', AnnClose '}',

Pattern Synonym binding

Constructors

PSB

Fields

psb_ext :: XPSB idL idR

Post renaming, FVs. See Note [Bind free vars]

psb_id :: Located (IdP idL)

Name of the pattern synonym

psb_args :: HsPatSynDetails (Located (IdP idR))

Formal parameter names

psb_def :: LPat idR

Right-hand side

psb_dir :: HsPatSynDir idR

Directionality

XPatSynBind (XXPatSynBind idL idR)

pprLHsBinds :: (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR)) => LHsBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

pprLHsBindsForUser :: (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR), OutputableBndrId (GhcPass id2)) => LHsBindsLR (GhcPass idL) (GhcPass idR) -> [LSig (GhcPass id2)] -> [SDoc]

pprDeclList :: [SDoc] -> SDoc

emptyLocalBinds :: HsLocalBindsLR (GhcPass a) (GhcPass b)

isEmptyLocalBindsTc :: HsLocalBindsLR (GhcPass a) GhcTc -> Bool

isEmptyLocalBindsPR :: HsLocalBindsLR (GhcPass a) (GhcPass b) -> Bool

eqEmptyLocalBinds :: HsLocalBindsLR a b -> Bool

isEmptyValBinds :: HsValBindsLR (GhcPass a) (GhcPass b) -> Bool

emptyValBindsIn :: HsValBindsLR (GhcPass a) (GhcPass b)

emptyValBindsOut :: HsValBindsLR (GhcPass a) (GhcPass b)

emptyLHsBinds :: LHsBindsLR idL idR

isEmptyLHsBinds :: LHsBindsLR idL idR -> Bool

plusHsValBinds :: HsValBinds (GhcPass a) -> HsValBinds (GhcPass a) -> HsValBinds (GhcPass a)

ppr_monobind :: (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR)) => HsBindLR (GhcPass idL) (GhcPass idR) -> SDoc

pprTicks :: SDoc -> SDoc -> SDoc

data HsIPBinds id

Haskell Implicit Parameter Bindings

Constructors

IPBinds (XIPBinds id) [LIPBind id]

XHsIPBinds (XXHsIPBinds id)

isEmptyIPBindsPR :: HsIPBinds (GhcPass p) -> Bool

isEmptyIPBindsTc :: HsIPBinds GhcTc -> Bool

type LIPBind id = Located (IPBind id)

Located Implicit Parameter Binding

May have AnnKeywordId : AnnSemi when in a list

data IPBind id

Implicit parameter bindings.

These bindings start off as (Left "x") in the parser and stay that way until after type-checking when they are replaced with (Right d), where "d" is the name of the dictionary holding the evidence for the implicit parameter.

AnnKeywordId : AnnEqual

Constructors

IPBind (XCIPBind id) (Either (Located HsIPName) (IdP id)) (LHsExpr id)

XIPBind (XXIPBind id)

type LSig pass = Located (Sig pass)

Located Signature

data Sig pass

Signatures and pragmas

Constructors

TypeSig (XTypeSig pass) [Located (IdP pass)] (LHsSigWcType pass)
An ordinary type signature

f :: Num a => a -> a

After renaming, this list of Names contains the named and unnamed wildcards brought into scope by this signature. For a signature _ -> _a -> Bool, the renamer will give the unnamed wildcard _ a freshly generated name, e.g. _w. _w and the named wildcard _a are then both replaced with fresh meta vars in the type. Their names are stored in the type signature that brought them into scope, in this third field to be more specific.

AnnKeywordId : AnnDcolon, AnnComma

PatSynSig (XPatSynSig pass) [Located (IdP pass)] (LHsSigType pass)
A pattern synonym type signature

pattern Single :: () => (Show a) => a -> [a]

AnnKeywordId : AnnPattern, AnnDcolon,AnnForall AnnDot,AnnDarrow

ClassOpSig (XClassOpSig pass) Bool [Located (IdP pass)] (LHsSigType pass)
A signature for a class method False: ordinary class-method signature True: generic-default class method signature e.g. class C a where op :: a -> a -- Ordinary default op :: Eq a => a -> a -- Generic default No wildcards allowed here

AnnKeywordId : AnnDefault, AnnDcolon

IdSig (XIdSig pass) Id
A type signature in generated code, notably the code generated for record selectors. We simply record the desired Id itself, replete with its name, type and IdDetails. Otherwise it's just like a type signature: there should be an accompanying binding

FixSig (XFixSig pass) (FixitySig pass)
An ordinary fixity declaration

    infixl 8 ***

AnnKeywordId : AnnInfix, AnnVal

InlineSig (XInlineSig pass) (Located (IdP pass)) InlinePragma
An inline pragma

{#- INLINE f #-}

AnnKeywordId : AnnOpen '{-# INLINE' and '[', AnnClose,AnnOpen, AnnVal,AnnTilde, AnnClose

SpecSig (XSpecSig pass) (Located (IdP pass)) [LHsSigType pass] InlinePragma
A specialisation pragma

{-# SPECIALISE f :: Int -> Int #-}

AnnKeywordId : AnnOpen, AnnOpen '{-# SPECIALISE' and '[', AnnTilde, AnnVal, AnnClose ']' and '#-}', AnnDcolon

SpecInstSig (XSpecInstSig pass) SourceText (LHsSigType pass)
A specialisation pragma for instance declarations only

{-# SPECIALISE instance Eq [Int] #-}

(Class tys); should be a specialisation of the current instance declaration

AnnKeywordId : AnnOpen, AnnInstance,AnnClose

MinimalSig (XMinimalSig pass) SourceText (LBooleanFormula (Located (IdP pass)))
A minimal complete definition pragma

{-# MINIMAL a | (b, c | (d | e)) #-}

AnnKeywordId : AnnOpen, AnnVbar,AnnComma, AnnClose

SCCFunSig (XSCCFunSig pass) SourceText (Located (IdP pass)) (Maybe (Located StringLiteral))
A "set cost centre" pragma for declarations

{-# SCC funName #-}

or

{-# SCC funName "cost_centre_name" #-}

CompleteMatchSig (XCompleteMatchSig pass) SourceText (Located [Located (IdP pass)]) (Maybe (Located (IdP pass)))
A complete match pragma

{-# COMPLETE C, D [:: T] #-}

Used to inform the pattern match checker about additional complete matchings which, for example, arise from pattern synonym definitions.

XSig (XXSig pass)

type LFixitySig pass = Located (FixitySig pass)

Located Fixity Signature

data FixitySig pass

Fixity Signature

Constructors

FixitySig (XFixitySig pass) [Located (IdP pass)] Fixity

XFixitySig (XXFixitySig pass)

data TcSpecPrags

Type checker Specialisation Pragmas

TcSpecPrags conveys SPECIALISE pragmas from the type checker to the desugarer

Constructors

IsDefaultMethod
Super-specialised: a default method should be macro-expanded at every call site

SpecPrags [LTcSpecPrag]

type LTcSpecPrag = Located TcSpecPrag

Located Type checker Specification Pragmas

data TcSpecPrag

Type checker Specification Pragma

Constructors

SpecPrag Id HsWrapper InlinePragma
The Id to be specialised, a wrapper that specialises the polymorphic function, and inlining spec for the specialised function

noSpecPrags :: TcSpecPrags

hasSpecPrags :: TcSpecPrags -> Bool

isDefaultMethod :: TcSpecPrags -> Bool

isFixityLSig :: LSig name -> Bool

isTypeLSig :: LSig name -> Bool

isSpecLSig :: LSig name -> Bool

isSpecInstLSig :: LSig name -> Bool

isPragLSig :: LSig name -> Bool

isInlineLSig :: LSig name -> Bool

isMinimalLSig :: LSig name -> Bool

isSCCFunSig :: LSig name -> Bool

isCompleteMatchSig :: LSig name -> Bool

hsSigDoc :: Sig name -> SDoc

ppr_sig :: OutputableBndrId (GhcPass p) => Sig (GhcPass p) -> SDoc

pragBrackets :: SDoc -> SDoc

pragSrcBrackets :: SourceText -> String -> SDoc -> SDoc

Using SourceText in case the pragma was spelled differently or used mixed case

pprVarSig :: OutputableBndr id => [id] -> SDoc -> SDoc

pprSpec :: OutputableBndr id => id -> SDoc -> InlinePragma -> SDoc

pprTcSpecPrags :: TcSpecPrags -> SDoc

pprMinimalSig :: OutputableBndr name => LBooleanFormula (Located name) -> SDoc

type HsPatSynDetails arg = HsConDetails arg [RecordPatSynField arg]

Haskell Pattern Synonym Details

data RecordPatSynField a

Record Pattern Synonym Field

Constructors

RecordPatSynField

Fields

recordPatSynSelectorId :: a

recordPatSynPatVar :: a

data HsPatSynDir id

Haskell Pattern Synonym Direction

Constructors

Unidirectional

ImplicitBidirectional

ExplicitBidirectional (MatchGroup id (LHsExpr id))