HsDecls

Description

Abstract syntax of global declarations.

Definitions for: SynDecl and ConDecl, ClassDecl, InstDecl, DefaultDecl and ForeignDecl.

Toplevel declarations

data HsDecl p

A Haskell Declaration

Constructors

TyClD (XTyClD p) (TyClDecl p)
Type or Class Declaration

InstD (XInstD p) (InstDecl p)
Instance declaration

DerivD (XDerivD p) (DerivDecl p)
Deriving declaration

ValD (XValD p) (HsBind p)
Value declaration

SigD (XSigD p) (Sig p)
Signature declaration

DefD (XDefD p) (DefaultDecl p)
'default' declaration

ForD (XForD p) (ForeignDecl p)
Foreign declaration

WarningD (XWarningD p) (WarnDecls p)
Warning declaration

AnnD (XAnnD p) (AnnDecl p)
Annotation declaration

RuleD (XRuleD p) (RuleDecls p)
Rule declaration

SpliceD (XSpliceD p) (SpliceDecl p)
Splice declaration (Includes quasi-quotes)

DocD (XDocD p) DocDecl
Documentation comment declaration

RoleAnnotD (XRoleAnnotD p) (RoleAnnotDecl p)
Role annotation declaration

XHsDecl (XXHsDecl p)

type LHsDecl p

Arguments

= Located (HsDecl p)
When in a list this may have

AnnKeywordId : AnnSemi

data HsDataDefn pass

Haskell Data type Definition

Constructors

HsDataDefn
Declares a data type or newtype, giving its constructors data/newtype T a = constrs data/newtype instance T [a] = constrs

Fields

dd_ext :: XCHsDataDefn pass

dd_ND :: NewOrData

dd_ctxt :: LHsContext pass

Context

dd_cType :: Maybe (Located CType)

dd_kindSig :: Maybe (LHsKind pass)

Optional kind signature.

(Just k) for a GADT-style data, or data instance decl, with explicit kind sig

Always Nothing for H98-syntax decls

dd_cons :: [LConDecl pass]

Data constructors

For data T a = T1 | T2 a the LConDecls all have ConDeclH98. For data T a where { T1 :: T a } the LConDecls all have ConDeclGADT.

dd_derivs :: HsDeriving pass

Optional 'deriving' claues

XHsDataDefn (XXHsDataDefn pass)

type HsDeriving pass

Arguments

= Located [LHsDerivingClause pass]
The optional deriving clauses of a data declaration. Clauses is plural because one can specify multiple deriving clauses using the -XDerivingStrategies language extension.

The list of LHsDerivingClauses corresponds to exactly what the user requested to derive, in order. If no deriving clauses were specified, the list is empty.

Haskell Deriving clause

data HsDerivingClause pass

A single deriving clause of a data declaration.

AnnKeywordId : AnnDeriving, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

Constructors

HsDerivingClause

Fields

deriv_clause_ext :: XCHsDerivingClause pass

deriv_clause_strategy :: Maybe (LDerivStrategy pass)

The user-specified strategy (if any) to use when deriving deriv_clause_tys.

deriv_clause_tys :: Located [LHsSigType pass]

The types to derive.

It uses LHsSigTypes because, with -XGeneralizedNewtypeDeriving, we can mention type variables that aren't bound by the datatype, e.g.

data T b = ... deriving (C [a])

should produce a derived instance for C [a] (T b).

XHsDerivingClause (XXHsDerivingClause pass)

type LHsDerivingClause pass = Located (HsDerivingClause pass)

data NewOrData

Constructors

NewType
newtype Blah ...

DataType
data Blah ...

newOrDataToFlavour :: NewOrData -> TyConFlavour

Convert a NewOrData to a TyConFlavour

Class or type declarations

data TyClDecl pass

A type or class declaration.

Constructors

FamDecl
type/data family T :: *->*

AnnKeywordId : AnnType, AnnData, AnnFamily,AnnDcolon, AnnWhere,AnnOpenP, AnnDcolon,AnnCloseP, AnnEqual,AnnRarrow, AnnVbar

Fields

tcdFExt :: XFamDecl pass

tcdFam :: FamilyDecl pass

SynDecl
type declaration

AnnKeywordId : AnnType, AnnEqual,

Fields

tcdSExt :: XSynDecl pass

Post renameer, FVs

tcdLName :: Located (IdP pass)

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdRhs :: LHsType pass

DataDecl
data declaration

AnnKeywordId : AnnData, AnnFamily, AnnNewType, AnnNewType,AnnDcolon AnnWhere,

Fields

tcdDExt :: XDataDecl pass

Post renamer, CUSK flag, FVs

tcdLName :: Located (IdP pass)

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdDataDefn :: HsDataDefn pass

ClassDecl

AnnKeywordId : AnnClass, AnnWhere,AnnOpen, AnnClose

The tcdFDs will have AnnVbar, AnnComma AnnRarrow

Fields

tcdCExt :: XClassDecl pass

Post renamer, FVs

tcdCtxt :: LHsContext pass

Context...

tcdLName :: Located (IdP pass)

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdFDs :: [Located (FunDep (Located (IdP pass)))]

Functional deps

tcdSigs :: [LSig pass]

Methods' signatures

tcdMeths :: LHsBinds pass

Default methods

tcdATs :: [LFamilyDecl pass]

Associated types;

tcdATDefs :: [LTyFamDefltEqn pass]

Associated type defaults

tcdDocs :: [LDocDecl]

Haddock docs

XTyClDecl (XXTyClDecl pass)

type LTyClDecl pass = Located (TyClDecl pass)

Located Declaration of a Type or Class

data DataDeclRn

Constructors

DataDeclRn

Fields

tcdDataCusk :: Bool

does this have a CUSK?

tcdFVs :: NameSet

data TyClGroup pass

Type or Class Group

Constructors

TyClGroup

Fields

group_ext :: XCTyClGroup pass

group_tyclds :: [LTyClDecl pass]

group_roles :: [LRoleAnnotDecl pass]

group_instds :: [LInstDecl pass]

XTyClGroup (XXTyClGroup pass)

mkTyClGroup :: [LTyClDecl (GhcPass p)] -> [LInstDecl (GhcPass p)] -> TyClGroup (GhcPass p)

emptyTyClGroup :: TyClGroup (GhcPass p)

tyClGroupTyClDecls :: [TyClGroup pass] -> [LTyClDecl pass]

tyClGroupInstDecls :: [TyClGroup pass] -> [LInstDecl pass]

tyClGroupRoleDecls :: [TyClGroup pass] -> [LRoleAnnotDecl pass]

isClassDecl :: TyClDecl pass -> Bool

type class

isDataDecl :: TyClDecl pass -> Bool

True = argument is a data/newtype declaration.

isSynDecl :: TyClDecl pass -> Bool

type or type instance declaration

tcdName :: TyClDecl pass -> IdP pass

isFamilyDecl :: TyClDecl pass -> Bool

type/data family declaration

isTypeFamilyDecl :: TyClDecl pass -> Bool

type family declaration

isDataFamilyDecl :: TyClDecl pass -> Bool

data family declaration

isOpenTypeFamilyInfo :: FamilyInfo pass -> Bool

open type family info

isClosedTypeFamilyInfo :: FamilyInfo pass -> Bool

closed type family info

tyFamInstDeclName :: TyFamInstDecl pass -> IdP pass

tyFamInstDeclLName :: TyFamInstDecl pass -> Located (IdP pass)

countTyClDecls :: [TyClDecl pass] -> (Int, Int, Int, Int, Int)

pprTyClDeclFlavour :: TyClDecl (GhcPass p) -> SDoc

tyClDeclLName :: TyClDecl pass -> Located (IdP pass)

tyClDeclTyVars :: TyClDecl pass -> LHsQTyVars pass

hsDeclHasCusk :: TyClDecl GhcRn -> Bool

Does this declaration have a complete, user-supplied kind signature? See Note [Complete user-supplied kind signatures]

famDeclHasCusk

Arguments

:: Maybe Bool
if associated, does the enclosing class have a CUSK?

-> FamilyDecl pass

-> Bool

Does this family declaration have a complete, user-supplied kind signature?

data FamilyDecl pass

type Family Declaration

Constructors

FamilyDecl

Fields

fdExt :: XCFamilyDecl pass

fdInfo :: FamilyInfo pass

fdLName :: Located (IdP pass)

fdTyVars :: LHsQTyVars pass

fdFixity :: LexicalFixity

fdResultSig :: LFamilyResultSig pass

fdInjectivityAnn :: Maybe (LInjectivityAnn pass)

XFamilyDecl (XXFamilyDecl pass)

AnnKeywordId : AnnType, AnnData, AnnFamily, AnnWhere, AnnOpenP, AnnDcolon, AnnCloseP, AnnEqual, AnnRarrow, AnnVbar

type LFamilyDecl pass = Located (FamilyDecl pass)

Located type Family Declaration

Instance declarations

data InstDecl pass

Instance Declaration

Constructors

ClsInstD

Fields

cid_d_ext :: XClsInstD pass

cid_inst :: ClsInstDecl pass

DataFamInstD

Fields

dfid_ext :: XDataFamInstD pass

dfid_inst :: DataFamInstDecl pass

TyFamInstD

Fields

tfid_ext :: XTyFamInstD pass

tfid_inst :: TyFamInstDecl pass

XInstDecl (XXInstDecl pass)

type LInstDecl pass = Located (InstDecl pass)

Located Instance Declaration

data FamilyInfo pass

Constructors

DataFamily

OpenTypeFamily

ClosedTypeFamily (Maybe [LTyFamInstEqn pass])
Nothing if we're in an hs-boot file and the user said "type family Foo x where .."

newtype TyFamInstDecl pass

Type Family Instance Declaration

Constructors

TyFamInstDecl

AnnKeywordId : AnnType, AnnInstance,

Fields

tfid_eqn :: TyFamInstEqn pass

type LTyFamInstDecl pass = Located (TyFamInstDecl pass)

Located Type Family Instance Declaration

instDeclDataFamInsts :: [LInstDecl pass] -> [DataFamInstDecl pass]

newtype DataFamInstDecl pass

Data Family Instance Declaration

Constructors

DataFamInstDecl

AnnKeywordId : AnnData, AnnNewType,AnnInstance, AnnDcolon AnnWhere,AnnOpen, AnnClose

Fields

dfid_eqn :: FamInstEqn pass (HsDataDefn pass)

type LDataFamInstDecl pass = Located (DataFamInstDecl pass)

Located Data Family Instance Declaration

pprDataFamInstFlavour :: DataFamInstDecl (GhcPass p) -> SDoc

pprFamInstLHS :: OutputableBndrId (GhcPass p) => Located (IdP (GhcPass p)) -> HsTyPats (GhcPass p) -> LexicalFixity -> HsContext (GhcPass p) -> Maybe (LHsKind (GhcPass p)) -> SDoc

type FamInstEqn pass rhs

Arguments

= HsImplicitBndrs pass (FamEqn pass (HsTyPats pass) rhs)
Here, the pats are type patterns (with kind and type bndrs). See Note [Family instance declaration binders]

Family Instance Equation

type LFamInstEqn pass rhs = Located (FamInstEqn pass rhs)

Located Family Instance Equation

data FamEqn pass pats rhs

Family Equation

One equation in a type family instance declaration, data family instance declaration, or type family default. See Note [Type family instance declarations in HsSyn] See Note [Family instance declaration binders]

Constructors

FamEqn

AnnKeywordId : AnnEqual

Fields

feqn_ext :: XCFamEqn pass pats rhs

feqn_tycon :: Located (IdP pass)

feqn_pats :: pats

feqn_fixity :: LexicalFixity

Fixity used in the declaration

feqn_rhs :: rhs

XFamEqn (XXFamEqn pass pats rhs)

type TyFamInstEqn pass = FamInstEqn pass (LHsType pass)

Type Family Instance Equation

type LTyFamInstEqn pass

Arguments

= Located (TyFamInstEqn pass)
May have AnnKeywordId : AnnSemi when in a list

Located Type Family Instance Equation

type TyFamDefltEqn pass = FamEqn pass (LHsQTyVars pass) (LHsType pass)

Type Family Default Equation

type LTyFamDefltEqn pass = Located (TyFamDefltEqn pass)

Located Type Family Default Equation

type HsTyPats pass = [LHsType pass]

Haskell Type Patterns

type LClsInstDecl pass = Located (ClsInstDecl pass)

Located Class Instance Declaration

data ClsInstDecl pass

Class Instance Declaration

Constructors

ClsInstDecl

AnnKeywordId : AnnInstance, AnnWhere, AnnOpen,AnnClose,

Fields

cid_ext :: XCClsInstDecl pass

cid_poly_ty :: LHsSigType pass

cid_binds :: LHsBinds pass

cid_sigs :: [LSig pass]

cid_tyfam_insts :: [LTyFamInstDecl pass]

cid_datafam_insts :: [LDataFamInstDecl pass]

cid_overlap_mode :: Maybe (Located OverlapMode)

AnnKeywordId : AnnOpen, AnnClose,

XClsInstDecl (XXClsInstDecl pass)

Standalone deriving declarations

data DerivDecl pass

Deriving Declaration

Constructors

DerivDecl

Fields

deriv_ext :: XCDerivDecl pass

deriv_type :: LHsSigWcType pass

The instance type to derive.

It uses an LHsSigWcType because the context is allowed to be a single wildcard:

deriving instance _ => Eq (Foo a)

Which signifies that the context should be inferred.

deriv_strategy :: Maybe (LDerivStrategy pass)

deriv_overlap_mode :: Maybe (Located OverlapMode)

AnnKeywordId : AnnDeriving, AnnInstance, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

XDerivDecl (XXDerivDecl pass)

type LDerivDecl pass = Located (DerivDecl pass)

Located Deriving Declaration

Deriving strategies

data DerivStrategy pass

Which technique the user explicitly requested when deriving an instance.

Constructors

StockStrategy
GHC's "standard" strategy, which is to implement a custom instance for the data type. This only works for certain types that GHC knows about (e.g., Eq, Show, Functor when -XDeriveFunctor is enabled, etc.)

AnyclassStrategy
-XDeriveAnyClass

NewtypeStrategy
-XGeneralizedNewtypeDeriving

ViaStrategy (XViaStrategy pass)
-XDerivingVia

type LDerivStrategy pass = Located (DerivStrategy pass)

A Located DerivStrategy.

derivStrategyName :: DerivStrategy a -> SDoc

A short description of a DerivStrategy'.

RULE declarations

type LRuleDecls pass = Located (RuleDecls pass)

Located Rule Declarations

data RuleDecls pass

Rule Declarations

Constructors

HsRules

Fields

rds_ext :: XCRuleDecls pass

rds_src :: SourceText

rds_rules :: [LRuleDecl pass]

XRuleDecls (XXRuleDecls pass)

data RuleDecl pass

Rule Declaration

Constructors

HsRule (XHsRule pass) (Located (SourceText, RuleName)) Activation [LRuleBndr pass] (Located (HsExpr pass)) (Located (HsExpr pass))

AnnKeywordId : AnnOpen,AnnTilde, AnnVal, AnnClose, AnnForall,AnnDot, AnnEqual,

XRuleDecl (XXRuleDecl pass)

type LRuleDecl pass = Located (RuleDecl pass)

Located Rule Declaration

data HsRuleRn

Constructors

HsRuleRn NameSet NameSet

data RuleBndr pass

Rule Binder

Constructors

RuleBndr (XCRuleBndr pass) (Located (IdP pass))

RuleBndrSig (XRuleBndrSig pass) (Located (IdP pass)) (LHsSigWcType pass)

XRuleBndr (XXRuleBndr pass)

AnnKeywordId : AnnOpen, AnnDcolon,AnnClose

type LRuleBndr pass = Located (RuleBndr pass)

Located Rule Binder

collectRuleBndrSigTys :: [RuleBndr pass] -> [LHsSigWcType pass]

flattenRuleDecls :: [LRuleDecls pass] -> [LRuleDecl pass]

pprFullRuleName :: Located (SourceText, RuleName) -> SDoc

default declarations

data DefaultDecl pass

Default Declaration

Constructors

DefaultDecl (XCDefaultDecl pass) [LHsType pass]

AnnKeywordIds : AnnDefault, AnnOpen,AnnClose

XDefaultDecl (XXDefaultDecl pass)

type LDefaultDecl pass = Located (DefaultDecl pass)

Located Default Declaration

Template haskell declaration splice

data SpliceExplicitFlag

Constructors

ExplicitSplice
= $(f x y)

ImplicitSplice
= f x y, i.e. a naked top level expression

data SpliceDecl p

Splice Declaration

Constructors

SpliceDecl (XSpliceDecl p) (Located (HsSplice p)) SpliceExplicitFlag

XSpliceDecl (XXSpliceDecl p)

type LSpliceDecl pass = Located (SpliceDecl pass)

Located Splice Declaration

Foreign function interface declarations

data ForeignDecl pass

Foreign Declaration

Constructors

ForeignImport

Fields

fd_i_ext :: XForeignImport pass

fd_name :: Located (IdP pass)

fd_sig_ty :: LHsSigType pass

fd_fi :: ForeignImport

ForeignExport

AnnKeywordId : AnnForeign, AnnImport,AnnExport, AnnDcolon

Fields

fd_e_ext :: XForeignExport pass

fd_name :: Located (IdP pass)

fd_sig_ty :: LHsSigType pass

fd_fe :: ForeignExport

XForeignDecl (XXForeignDecl pass)

type LForeignDecl pass = Located (ForeignDecl pass)

Located Foreign Declaration

data ForeignImport

Constructors

CImport (Located CCallConv) (Located Safety) (Maybe Header) CImportSpec (Located SourceText)

data ForeignExport

Constructors

CExport (Located CExportSpec) (Located SourceText)

data CImportSpec

Constructors

CLabel CLabelString

CFunction CCallTarget

CWrapper

Data-constructor declarations

data ConDecl pass

data T b = forall a. Eq a => MkT a b
  MkT :: forall b a. Eq a => MkT a b

data T b where
     MkT1 :: Int -> T Int

data T = Int MkT Int
       | MkT2

data T a where
     Int MkT Int :: T Int

AnnKeywordIds : AnnOpen, AnnDotdot,AnnCLose, AnnEqual,AnnVbar, AnnDarrow,AnnDarrow, AnnForall,AnnDot

data Constructor Declaration

Constructors

ConDeclGADT

Fields

con_g_ext :: XConDeclGADT pass

con_names :: [Located (IdP pass)]

con_forall :: Located Bool

True = explicit forall False => hsq_explicit is empty

con_qvars :: LHsQTyVars pass

con_mb_cxt :: Maybe (LHsContext pass)

User-written context (if any)

con_args :: HsConDeclDetails pass

Arguments; never InfixCon

con_res_ty :: LHsType pass

Result type

con_doc :: Maybe LHsDocString

A possible Haddock comment.

ConDeclH98

Fields

con_ext :: XConDeclH98 pass

con_name :: Located (IdP pass)

con_forall :: Located Bool

True = explicit forall False => hsq_explicit is empty

con_ex_tvs :: [LHsTyVarBndr pass]

Existentials only

con_mb_cxt :: Maybe (LHsContext pass)

User-written context (if any)

con_args :: HsConDeclDetails pass

Arguments; never InfixCon

con_doc :: Maybe LHsDocString

A possible Haddock comment.

XConDecl (XXConDecl pass)

type LConDecl pass

Arguments

= Located (ConDecl pass)
May have AnnKeywordId : AnnSemi when in a GADT constructor list

Located data Constructor Declaration

type HsConDeclDetails pass = HsConDetails (LBangType pass) (Located [LConDeclField pass])

Haskell data Constructor Declaration Details

hsConDeclArgTys :: HsConDeclDetails pass -> [LBangType pass]

hsConDeclTheta :: Maybe (LHsContext pass) -> [LHsType pass]

getConNames :: ConDecl pass -> [Located (IdP pass)]

getConArgs :: ConDecl pass -> HsConDeclDetails pass

Document comments

data DocDecl

Documentation comment Declaration

Constructors

DocCommentNext HsDocString

DocCommentPrev HsDocString

DocCommentNamed String HsDocString

DocGroup Int HsDocString

type LDocDecl = Located DocDecl

Located Documentation comment Declaration

docDeclDoc :: DocDecl -> HsDocString

Deprecations

data WarnDecl pass

Warning pragma Declaration

Constructors

Warning (XWarning pass) [Located (IdP pass)] WarningTxt

XWarnDecl (XXWarnDecl pass)

type LWarnDecl pass = Located (WarnDecl pass)

Located Warning pragma Declaration

data WarnDecls pass

Warning pragma Declarations

Constructors

Warnings

Fields

wd_ext :: XWarnings pass

wd_src :: SourceText

wd_warnings :: [LWarnDecl pass]

XWarnDecls (XXWarnDecls pass)

type LWarnDecls pass = Located (WarnDecls pass)

Located Warning Declarations

Annotations

data AnnDecl pass

Annotation Declaration

Constructors

HsAnnotation (XHsAnnotation pass) SourceText (AnnProvenance (IdP pass)) (Located (HsExpr pass))

AnnKeywordId : AnnOpen, AnnType AnnModule AnnClose

XAnnDecl (XXAnnDecl pass)

type LAnnDecl pass = Located (AnnDecl pass)

Located Annotation Declaration

data AnnProvenance name

Annotation Provenance

Constructors

ValueAnnProvenance (Located name)

TypeAnnProvenance (Located name)

ModuleAnnProvenance

annProvenanceName_maybe :: AnnProvenance name -> Maybe name

Role annotations

data RoleAnnotDecl pass

Role Annotation Declaration

Constructors

RoleAnnotDecl (XCRoleAnnotDecl pass) (Located (IdP pass)) [Located (Maybe Role)]

AnnKeywordId : AnnType, AnnRole

XRoleAnnotDecl (XXRoleAnnotDecl pass)

type LRoleAnnotDecl pass = Located (RoleAnnotDecl pass)

Located Role Annotation Declaration

roleAnnotDeclName :: RoleAnnotDecl pass -> IdP pass

Injective type families

data FamilyResultSig pass

type Family Result Signature

Constructors

NoSig (XNoSig pass)

AnnKeywordId :

KindSig (XCKindSig pass) (LHsKind pass)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP

TyVarSig (XTyVarSig pass) (LHsTyVarBndr pass)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP, AnnEqual

XFamilyResultSig (XXFamilyResultSig pass)

type LFamilyResultSig pass = Located (FamilyResultSig pass)

Located type Family Result Signature

data InjectivityAnn pass

If the user supplied an injectivity annotation it is represented using InjectivityAnn. At the moment this is a single injectivity condition - see Note [Injectivity annotation]. `Located name` stores the LHS of injectivity condition. `[Located name]` stores the RHS of injectivity condition. Example:

type family Foo a b c = r | r -> a c where ...

This will be represented as "InjectivityAnn r [a, c]"

Constructors

InjectivityAnn (Located (IdP pass)) [Located (IdP pass)]

AnnKeywordId : AnnRarrow, AnnVbar

type LInjectivityAnn pass = Located (InjectivityAnn pass)

Located Injectivity Annotation

resultVariableName :: FamilyResultSig a -> Maybe (IdP a)

Maybe return name of the result type variable

Grouping

data HsGroup p

Haskell Group

A HsDecl is categorised into a HsGroup before being fed to the renamer.

Constructors

HsGroup

Fields

hs_ext :: XCHsGroup p

hs_valds :: HsValBinds p

hs_splcds :: [LSpliceDecl p]

hs_tyclds :: [TyClGroup p]

hs_derivds :: [LDerivDecl p]

hs_fixds :: [LFixitySig p]

hs_defds :: [LDefaultDecl p]

hs_fords :: [LForeignDecl p]

hs_warnds :: [LWarnDecls p]

hs_annds :: [LAnnDecl p]

hs_ruleds :: [LRuleDecls p]

hs_docs :: [LDocDecl]

XHsGroup (XXHsGroup p)

emptyRdrGroup :: HsGroup (GhcPass p)

emptyRnGroup :: HsGroup (GhcPass p)

appendGroups :: HsGroup (GhcPass p) -> HsGroup (GhcPass p) -> HsGroup (GhcPass p)

hsGroupInstDecls :: HsGroup id -> [LInstDecl id]