HsExpr

Description

Abstract Haskell syntax for expressions.

Expressions proper

type LHsExpr p

Arguments

= Located (HsExpr p)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Expression

type PostTcExpr = HsExpr GhcTc

Post-Type checking Expression

PostTcExpr is an evidence expression attached to the syntax tree by the type checker (c.f. postTcType).

type PostTcTable = [(Name, PostTcExpr)]

Post-Type checking Table

We use a PostTcTable where there are a bunch of pieces of evidence, more than is convenient to keep individually.

data SyntaxExpr p

Syntax Expression

SyntaxExpr is like PostTcExpr, but it's filled in a little earlier, by the renamer. It's used for rebindable syntax.

E.g. (>>=) is filled in before the renamer by the appropriate Name for (>>=), and then instantiated by the type checker with its type args etc

This should desugar to

syn_res_wrap $ syn_expr (syn_arg_wraps[0] arg0)
                        (syn_arg_wraps[1] arg1) ...

where the actual arguments come from elsewhere in the AST. This could be defined using GhcPass p and such, but it's harder to get it all to work out that way. (noSyntaxExpr is hard to write, for example.)

Constructors

SyntaxExpr

Fields

syn_expr :: HsExpr p

syn_arg_wraps :: [HsWrapper]

syn_res_wrap :: HsWrapper

noExpr :: HsExpr (GhcPass p)

This is used for rebindable-syntax pieces that are too polymorphic for tcSyntaxOp (trS_fmap and the mzip in ParStmt)

noSyntaxExpr :: SyntaxExpr (GhcPass p)

mkRnSyntaxExpr :: Name -> SyntaxExpr GhcRn

Make a 'SyntaxExpr Name' (the "rn" is because this is used in the renamer), missing its HsWrappers.

type CmdSyntaxTable p = [(Name, HsExpr p)]

Command Syntax Table (for Arrow syntax)

data UnboundVar

An unbound variable; used for treating out-of-scope variables as expression holes

Constructors

OutOfScope OccName GlobalRdrEnv
An (unqualified) out-of-scope variable, together with the GlobalRdrEnv with respect to which it is unbound

TrueExprHole OccName
A "true" expression hole (_ or _x)

unboundVarOcc :: UnboundVar -> OccName

data HsExpr p

A Haskell expression.

Constructors

HsVar (XVar p) (Located (IdP p))
Variable

HsUnboundVar (XUnboundVar p) UnboundVar
Unbound variable; also used for "holes" (_ or _x). Turned from HsVar to HsUnboundVar by the renamer, when it finds an out-of-scope variable or hole. Turned into HsVar by type checker, to support deferred type errors.

HsConLikeOut (XConLikeOut p) ConLike
After typechecker only; must be different HsVar for pretty printing

HsRecFld (XRecFld p) (AmbiguousFieldOcc p)
Variable pointing to record selector Not in use after typechecking

HsOverLabel (XOverLabel p) (Maybe (IdP p)) FastString
Overloaded label (Note [Overloaded labels] in GHC.OverloadedLabels) Just id means RebindableSyntax is in use, and gives the id of the in-scope fromLabel. NB: Not in use after typechecking

HsIPVar (XIPVar p) HsIPName
Implicit parameter (not in use after typechecking)

HsOverLit (XOverLitE p) (HsOverLit p)
Overloaded literals

HsLit (XLitE p) (HsLit p)
Simple (non-overloaded) literals

HsLam (XLam p) (MatchGroup p (LHsExpr p))
Lambda abstraction. Currently always a single match

AnnKeywordId : AnnLam, AnnRarrow,

HsLamCase (XLamCase p) (MatchGroup p (LHsExpr p))
Lambda-case

AnnKeywordId : AnnLam, AnnCase,AnnOpen, AnnClose

HsApp (XApp p) (LHsExpr p) (LHsExpr p)
Application

HsAppType (XAppTypeE p) (LHsExpr p)
Visible type application

Explicit type argument; e.g f @Int x y NB: Has wildcards, but no implicit quantification

AnnKeywordId : AnnAt,

OpApp (XOpApp p) (LHsExpr p) (LHsExpr p) (LHsExpr p)
Operator applications: NB Bracketed ops such as (+) come out as Vars.

NegApp (XNegApp p) (LHsExpr p) (SyntaxExpr p)
Negation operator. Contains the negated expression and the name of negate

AnnKeywordId : AnnMinus

HsPar (XPar p) (LHsExpr p)
Parenthesised expr; see Note [Parens in HsSyn]

SectionL (XSectionL p) (LHsExpr p) (LHsExpr p)

SectionR (XSectionR p) (LHsExpr p) (LHsExpr p)

ExplicitTuple (XExplicitTuple p) [LHsTupArg p] Boxity
Used for explicit tuples and sections thereof

AnnKeywordId : AnnOpen, AnnClose

ExplicitSum (XExplicitSum p) ConTag Arity (LHsExpr p)
Used for unboxed sum types

AnnKeywordId : AnnOpen '(#', AnnVbar, AnnClose '#)',

There will be multiple AnnVbar, (1 - alternative) before the expression, (arity - alternative) after it

HsCase (XCase p) (LHsExpr p) (MatchGroup p (LHsExpr p))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsIf (XIf p) (Maybe (SyntaxExpr p)) (LHsExpr p) (LHsExpr p) (LHsExpr p)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsMultiIf (XMultiIf p) [LGRHS p (LHsExpr p)]
Multi-way if

AnnKeywordId : AnnIf AnnOpen,AnnClose,

HsLet (XLet p) (LHsLocalBinds p) (LHsExpr p)
let(rec)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsDo (XDo p) (HsStmtContext Name) (Located [ExprLStmt p])

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

ExplicitList (XExplicitList p) (Maybe (SyntaxExpr p)) [LHsExpr p]
Syntactic list: [a,b,c,...]

AnnKeywordId : AnnOpen '[', AnnClose ']'

RecordCon
Record construction

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rcon_ext :: XRecordCon p

rcon_con_name :: Located (IdP p)

rcon_flds :: HsRecordBinds p

RecordUpd
Record update

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rupd_ext :: XRecordUpd p

rupd_expr :: LHsExpr p

rupd_flds :: [LHsRecUpdField p]

ExprWithTySig (XExprWithTySig p) (LHsExpr p)
Expression with an explicit type signature. e :: type

AnnKeywordId : AnnDcolon

ArithSeq (XArithSeq p) (Maybe (SyntaxExpr p)) (ArithSeqInfo p)
Arithmetic sequence

AnnKeywordId : AnnOpen '[', AnnComma,AnnDotdot, AnnClose ']'

HsSCC (XSCC p) SourceText StringLiteral (LHsExpr p)

HsCoreAnn (XCoreAnn p) SourceText StringLiteral (LHsExpr p)

AnnKeywordId : AnnOpen '{-# CORE', AnnVal, AnnClose '#-}'

HsBracket (XBracket p) (HsBracket p)

AnnKeywordId : AnnOpen, AnnOpenE,AnnOpenEQ, AnnClose,AnnCloseQ

HsRnBracketOut (XRnBracketOut p) (HsBracket GhcRn) [PendingRnSplice]

HsTcBracketOut (XTcBracketOut p) (HsBracket GhcRn) [PendingTcSplice]

HsSpliceE (XSpliceE p) (HsSplice p)

AnnKeywordId : AnnOpen, AnnClose

HsProc (XProc p) (LPat p) (LHsCmdTop p)
proc notation for Arrows

AnnKeywordId : AnnProc, AnnRarrow

HsStatic (XStatic p) (LHsExpr p)

AnnKeywordId : AnnStatic,

HsArrApp (XArrApp p) (LHsExpr p) (LHsExpr p) HsArrAppType Bool

AnnKeywordId : Annlarrowtail, Annrarrowtail,AnnLarrowtail, AnnRarrowtail

HsArrForm (XArrForm p) (LHsExpr p) (Maybe Fixity) [LHsCmdTop p]

AnnKeywordId : AnnOpenB '(|', AnnCloseB '|)'

HsTick (XTick p) (Tickish (IdP p)) (LHsExpr p)

HsBinTick (XBinTick p) Int Int (LHsExpr p)

HsTickPragma (XTickPragma p) SourceText (StringLiteral, (Int, Int), (Int, Int)) ((SourceText, SourceText), (SourceText, SourceText)) (LHsExpr p)

AnnKeywordId : AnnOpen, AnnOpen '{-# GENERATED', AnnVal,AnnVal, AnnColon,AnnVal, AnnMinus, AnnVal,AnnColon, AnnVal, AnnClose '#-}'

EWildPat (XEWildPat p)

EAsPat (XEAsPat p) (Located (IdP p)) (LHsExpr p)

AnnKeywordId : AnnAt

EViewPat (XEViewPat p) (LHsExpr p) (LHsExpr p)

AnnKeywordId : AnnRarrow

ELazyPat (XELazyPat p) (LHsExpr p)

AnnKeywordId : AnnTilde

HsWrap (XWrap p) HsWrapper (HsExpr p)

XExpr (XXExpr p)

data RecordConTc

Extra data fields for a RecordCon, added by the type checker

Constructors

RecordConTc

Fields

rcon_con_like :: ConLike

rcon_con_expr :: PostTcExpr

data RecordUpdTc

Extra data fields for a RecordUpd, added by the type checker

Constructors

RecordUpdTc

Fields

rupd_cons :: [ConLike]

rupd_in_tys :: [Type]

rupd_out_tys :: [Type]

rupd_wrap :: HsWrapper

type LHsTupArg id = Located (HsTupArg id)

Located Haskell Tuple Argument

HsTupArg is used for tuple sections (,a,) is represented by ExplicitTuple [Missing ty1, Present a, Missing ty3] Which in turn stands for (x:ty1 y:ty2. (x,a,y))

data HsTupArg id

AnnKeywordId : AnnComma

Haskell Tuple Argument

Constructors

Present (XPresent id) (LHsExpr id)
The argument

Missing (XMissing id)
The argument is missing, but this is its type

XTupArg (XXTupArg id)
Note [Trees that Grow] extension point

tupArgPresent :: LHsTupArg id -> Bool

pprLExpr :: OutputableBndrId (GhcPass p) => LHsExpr (GhcPass p) -> SDoc

pprExpr :: OutputableBndrId (GhcPass p) => HsExpr (GhcPass p) -> SDoc

isQuietHsExpr :: HsExpr id -> Bool

pprBinds :: (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR)) => HsLocalBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

ppr_lexpr :: OutputableBndrId (GhcPass p) => LHsExpr (GhcPass p) -> SDoc

ppr_expr :: forall p. OutputableBndrId (GhcPass p) => HsExpr (GhcPass p) -> SDoc

ppr_apps :: OutputableBndrId (GhcPass p) => HsExpr (GhcPass p) -> [Either (LHsExpr (GhcPass p)) (XAppTypeE (GhcPass p))] -> SDoc

pprExternalSrcLoc :: (StringLiteral, (Int, Int), (Int, Int)) -> SDoc

pprDebugParendExpr :: OutputableBndrId (GhcPass p) => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprParendLExpr :: OutputableBndrId (GhcPass p) => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprParendExpr :: OutputableBndrId (GhcPass p) => PprPrec -> HsExpr (GhcPass p) -> SDoc

hsExprNeedsParens :: PprPrec -> HsExpr p -> Bool

hsExprNeedsParens p e returns True if the expression e needs parentheses under precedence p.

parenthesizeHsExpr :: PprPrec -> LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

parenthesizeHsExpr p e checks if hsExprNeedsParens p e is true, and if so, surrounds e with an HsPar. Otherwise, it simply returns e.

isAtomicHsExpr :: HsExpr id -> Bool

type LHsCmd id = Located (HsCmd id)

Located Haskell Command (for arrow syntax)

data HsCmd id

Haskell Command (e.g. a "statement" in an Arrow proc block)

Constructors

HsCmdArrApp (XCmdArrApp id) (LHsExpr id) (LHsExpr id) HsArrAppType Bool

AnnKeywordId : Annlarrowtail, Annrarrowtail,AnnLarrowtail, AnnRarrowtail

HsCmdArrForm (XCmdArrForm id) (LHsExpr id) LexicalFixity (Maybe Fixity) [LHsCmdTop id]

AnnKeywordId : AnnOpenB '(|', AnnCloseB '|)'

HsCmdApp (XCmdApp id) (LHsCmd id) (LHsExpr id)

HsCmdLam (XCmdLam id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnLam, AnnRarrow,

HsCmdPar (XCmdPar id) (LHsCmd id)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsCmdCase (XCmdCase id) (LHsExpr id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsCmdIf (XCmdIf id) (Maybe (SyntaxExpr id)) (LHsExpr id) (LHsCmd id) (LHsCmd id)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsCmdLet (XCmdLet id) (LHsLocalBinds id) (LHsCmd id)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsCmdDo (XCmdDo id) (Located [CmdLStmt id])

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

HsCmdWrap (XCmdWrap id) HsWrapper (HsCmd id)

XCmd (XXCmd id)

data HsArrAppType

Haskell Array Application Type

Constructors

HsHigherOrderApp

HsFirstOrderApp

type LHsCmdTop p = Located (HsCmdTop p)

Top-level command, introducing a new arrow. This may occur inside a proc (where the stack is empty) or as an argument of a command-forming operator.

Located Haskell Top-level Command

data HsCmdTop p

Haskell Top-level Command

Constructors

HsCmdTop (XCmdTop p) (LHsCmd p)

XCmdTop (XXCmdTop p)

data CmdTopTc

Constructors

CmdTopTc Type Type (CmdSyntaxTable GhcTc)

pprLCmd :: OutputableBndrId (GhcPass p) => LHsCmd (GhcPass p) -> SDoc

pprCmd :: OutputableBndrId (GhcPass p) => HsCmd (GhcPass p) -> SDoc

isQuietHsCmd :: HsCmd id -> Bool

ppr_lcmd :: OutputableBndrId (GhcPass p) => LHsCmd (GhcPass p) -> SDoc

ppr_cmd :: forall p. OutputableBndrId (GhcPass p) => HsCmd (GhcPass p) -> SDoc

pprCmdArg :: OutputableBndrId (GhcPass p) => HsCmdTop (GhcPass p) -> SDoc

type HsRecordBinds p = HsRecFields p (LHsExpr p)

Haskell Record Bindings

data MatchGroup p body

Constructors

MG

Fields

mg_ext :: XMG p body

mg_alts :: Located [LMatch p body]

mg_origin :: Origin

XMatchGroup (XXMatchGroup p body)

data MatchGroupTc

Constructors

MatchGroupTc

Fields

mg_arg_tys :: [Type]

mg_res_ty :: Type

type LMatch id body = Located (Match id body)

Located Match

May have AnnKeywordId : AnnSemi when in a list

data Match p body

Constructors

Match

Fields

m_ext :: XCMatch p body

m_ctxt :: HsMatchContext (NameOrRdrName (IdP p))

m_pats :: [LPat p]

m_grhss :: GRHSs p body

XMatch (XXMatch p body)

isInfixMatch :: Match id body -> Bool

isEmptyMatchGroup :: MatchGroup id body -> Bool

isSingletonMatchGroup :: [LMatch id body] -> Bool

Is there only one RHS in this list of matches?

matchGroupArity :: MatchGroup id body -> Arity

hsLMatchPats :: LMatch id body -> [LPat id]

data GRHSs p body

Guarded Right-Hand Sides

GRHSs are used both for pattern bindings and for Matches

AnnKeywordId : AnnVbar, AnnEqual,AnnWhere, AnnOpen,AnnClose AnnRarrow,AnnSemi

Constructors

GRHSs

Fields

grhssExt :: XCGRHSs p body

grhssGRHSs :: [LGRHS p body]

Guarded RHSs

grhssLocalBinds :: LHsLocalBinds p

The where clause

XGRHSs (XXGRHSs p body)

type LGRHS id body = Located (GRHS id body)

Located Guarded Right-Hand Side

data GRHS p body

Guarded Right Hand Side.

Constructors

GRHS (XCGRHS p body) [GuardLStmt p] body

XGRHS (XXGRHS p body)

pprMatches :: (OutputableBndrId (GhcPass idR), Outputable body) => MatchGroup (GhcPass idR) body -> SDoc

pprFunBind :: (OutputableBndrId (GhcPass idR), Outputable body) => MatchGroup (GhcPass idR) body -> SDoc

pprPatBind :: forall bndr p body. (OutputableBndrId (GhcPass bndr), OutputableBndrId (GhcPass p), Outputable body) => LPat (GhcPass bndr) -> GRHSs (GhcPass p) body -> SDoc

pprMatch :: (OutputableBndrId (GhcPass idR), Outputable body) => Match (GhcPass idR) body -> SDoc

pprGRHSs :: (OutputableBndrId (GhcPass idR), Outputable body) => HsMatchContext idL -> GRHSs (GhcPass idR) body -> SDoc

pprGRHS :: (OutputableBndrId (GhcPass idR), Outputable body) => HsMatchContext idL -> GRHS (GhcPass idR) body -> SDoc

pp_rhs :: Outputable body => HsMatchContext idL -> body -> SDoc

type LStmt id body = Located (StmtLR id id body)

Located do block Statement

type LStmtLR idL idR body = Located (StmtLR idL idR body)

Located Statement with separate Left and Right id's

type Stmt id body = StmtLR id id body

do block Statement

type CmdLStmt id = LStmt id (LHsCmd id)

Command Located Statement

type CmdStmt id = Stmt id (LHsCmd id)

Command Statement

type ExprLStmt id = LStmt id (LHsExpr id)

Expression Located Statement

type ExprStmt id = Stmt id (LHsExpr id)

Expression Statement

type GuardLStmt id = LStmt id (LHsExpr id)

Guard Located Statement

type GuardStmt id = Stmt id (LHsExpr id)

Guard Statement

type GhciLStmt id = LStmt id (LHsExpr id)

Ghci Located Statement

type GhciStmt id = Stmt id (LHsExpr id)

Ghci Statement

data StmtLR idL idR body

API Annotations when in qualifier lists or guards - AnnKeywordId : AnnVbar, AnnComma,AnnThen, AnnBy,AnnBy, AnnGroup,AnnUsing

Constructors

LastStmt (XLastStmt idL idR body) body Bool (SyntaxExpr idR)

BindStmt (XBindStmt idL idR body) (LPat idL) body (SyntaxExpr idR) (SyntaxExpr idR)

ApplicativeStmt (XApplicativeStmt idL idR body) [(SyntaxExpr idR, ApplicativeArg idL)] (Maybe (SyntaxExpr idR))
ApplicativeStmt represents an applicative expression built with $ and *. It is generated by the renamer, and is desugared into the appropriate applicative expression by the desugarer, but it is intended to be invisible in error messages.

For full details, see Note [ApplicativeDo] in RnExpr

BodyStmt (XBodyStmt idL idR body) body (SyntaxExpr idR) (SyntaxExpr idR)

LetStmt (XLetStmt idL idR body) (LHsLocalBindsLR idL idR)

AnnKeywordId : AnnLet AnnOpen '{',AnnClose '}',

ParStmt (XParStmt idL idR body) [ParStmtBlock idL idR] (HsExpr idR) (SyntaxExpr idR)

TransStmt

Fields

trS_ext :: XTransStmt idL idR body

trS_form :: TransForm

trS_stmts :: [ExprLStmt idL]

trS_bndrs :: [(IdP idR, IdP idR)]

trS_using :: LHsExpr idR

trS_by :: Maybe (LHsExpr idR)

trS_ret :: SyntaxExpr idR

trS_bind :: SyntaxExpr idR

trS_fmap :: HsExpr idR

RecStmt

AnnKeywordId : AnnRec

Fields

recS_ext :: XRecStmt idL idR body

recS_stmts :: [LStmtLR idL idR body]

recS_later_ids :: [IdP idR]

recS_rec_ids :: [IdP idR]

recS_bind_fn :: SyntaxExpr idR

recS_ret_fn :: SyntaxExpr idR

recS_mfix_fn :: SyntaxExpr idR

XStmtLR (XXStmtLR idL idR body)

data RecStmtTc

Constructors

RecStmtTc

Fields

recS_bind_ty :: Type

recS_later_rets :: [PostTcExpr]

recS_rec_rets :: [PostTcExpr]

recS_ret_ty :: Type

data TransForm

Constructors

ThenForm

GroupForm

data ParStmtBlock idL idR

Parenthesised Statement Block

Constructors

ParStmtBlock (XParStmtBlock idL idR) [ExprLStmt idL] [IdP idR] (SyntaxExpr idR)

XParStmtBlock (XXParStmtBlock idL idR)

data ApplicativeArg idL

Applicative Argument

Constructors

ApplicativeArgOne (XApplicativeArgOne idL) (LPat idL) (LHsExpr idL) Bool

ApplicativeArgMany (XApplicativeArgMany idL) [ExprLStmt idL] (HsExpr idL) (LPat idL)

XApplicativeArg (XXApplicativeArg idL)

pprStmt :: forall idL idR body. (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR), Outputable body) => StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc

pprTransformStmt :: OutputableBndrId (GhcPass p) => [IdP (GhcPass p)] -> LHsExpr (GhcPass p) -> Maybe (LHsExpr (GhcPass p)) -> SDoc

pprTransStmt :: Outputable body => Maybe body -> body -> TransForm -> SDoc

pprBy :: Outputable body => Maybe body -> SDoc

pprDo :: (OutputableBndrId (GhcPass p), Outputable body) => HsStmtContext any -> [LStmt (GhcPass p) body] -> SDoc

ppr_do_stmts :: (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR), Outputable body) => [LStmtLR (GhcPass idL) (GhcPass idR) body] -> SDoc

pprComp :: (OutputableBndrId (GhcPass p), Outputable body) => [LStmt (GhcPass p) body] -> SDoc

pprQuals :: (OutputableBndrId (GhcPass p), Outputable body) => [LStmt (GhcPass p) body] -> SDoc

data HsSplice id

Haskell Splice

Constructors

HsTypedSplice (XTypedSplice id) SpliceDecoration (IdP id) (LHsExpr id)

HsUntypedSplice (XUntypedSplice id) SpliceDecoration (IdP id) (LHsExpr id)

HsQuasiQuote (XQuasiQuote id) (IdP id) (IdP id) SrcSpan FastString

HsSpliced (XSpliced id) ThModFinalizers (HsSplicedThing id)

XSplice (XXSplice id)

data SpliceDecoration

A splice can appear with various decorations wrapped around it. This data type captures explicitly how it was originally written, for use in the pretty printer.

Constructors

HasParens
$( splice ) or $$( splice )

HasDollar
$splice or $$splice

NoParens
bare splice

isTypedSplice :: HsSplice id -> Bool

newtype ThModFinalizers

Finalizers produced by a splice with addModFinalizer

See Note [Delaying modFinalizers in untyped splices] in RnSplice. For how this is used.

Constructors

ThModFinalizers [ForeignRef (Q ())]

data HsSplicedThing id

Haskell Spliced Thing

Values that can result from running a splice.

Constructors

HsSplicedExpr (HsExpr id)
Haskell Spliced Expression

HsSplicedTy (HsType id)
Haskell Spliced Type

HsSplicedPat (Pat id)
Haskell Spliced Pattern

type SplicePointName = Name

data PendingRnSplice

Pending Renamer Splice

Constructors

PendingRnSplice UntypedSpliceFlavour SplicePointName (LHsExpr GhcRn)

data UntypedSpliceFlavour

Constructors

UntypedExpSplice

UntypedPatSplice

UntypedTypeSplice

UntypedDeclSplice

data PendingTcSplice

Pending Type-checker Splice

Constructors

PendingTcSplice SplicePointName (LHsExpr GhcTc)

pprPendingSplice :: OutputableBndrId (GhcPass p) => SplicePointName -> LHsExpr (GhcPass p) -> SDoc

pprSpliceDecl :: OutputableBndrId (GhcPass p) => HsSplice (GhcPass p) -> SpliceExplicitFlag -> SDoc

ppr_splice_decl :: OutputableBndrId (GhcPass p) => HsSplice (GhcPass p) -> SDoc

pprSplice :: OutputableBndrId (GhcPass p) => HsSplice (GhcPass p) -> SDoc

ppr_quasi :: OutputableBndr p => p -> p -> FastString -> SDoc

ppr_splice :: OutputableBndrId (GhcPass p) => SDoc -> IdP (GhcPass p) -> LHsExpr (GhcPass p) -> SDoc -> SDoc

data HsBracket p

Haskell Bracket

Constructors

ExpBr (XExpBr p) (LHsExpr p)

PatBr (XPatBr p) (LPat p)

DecBrL (XDecBrL p) [LHsDecl p]

DecBrG (XDecBrG p) (HsGroup p)

TypBr (XTypBr p) (LHsType p)

VarBr (XVarBr p) Bool (IdP p)

TExpBr (XTExpBr p) (LHsExpr p)

XBracket (XXBracket p)

isTypedBracket :: HsBracket id -> Bool

pprHsBracket :: OutputableBndrId (GhcPass p) => HsBracket (GhcPass p) -> SDoc

thBrackets :: SDoc -> SDoc -> SDoc

thTyBrackets :: SDoc -> SDoc

data ArithSeqInfo id

Arithmetic Sequence Information

Constructors

From (LHsExpr id)

FromThen (LHsExpr id) (LHsExpr id)

FromTo (LHsExpr id) (LHsExpr id)

FromThenTo (LHsExpr id) (LHsExpr id) (LHsExpr id)

pp_dotdot :: SDoc

data HsMatchContext id

Haskell Match Context

Context of a pattern match. This is more subtle than it would seem. See Note [Varieties of pattern matches].

Constructors

FunRhs
A pattern matching on an argument of a function binding

Fields

mc_fun :: Located id

function binder of f

mc_fixity :: LexicalFixity

fixing of f

mc_strictness :: SrcStrictness

was f banged? See Note [FunBind vs PatBind]

LambdaExpr
Patterns of a lambda

CaseAlt
Patterns and guards on a case alternative

IfAlt
Guards of a multi-way if alternative

ProcExpr
Patterns of a proc

PatBindRhs
A pattern binding eg [y] <- e = e

PatBindGuards
Guards of pattern bindings, e.g., (Just b) | Just _ <- x = e | otherwise = e'

RecUpd
Record update [used only in DsExpr to tell matchWrapper what sort of runtime error message to generate]

StmtCtxt (HsStmtContext id)
Pattern of a do-stmt, list comprehension, pattern guard, etc

ThPatSplice
A Template Haskell pattern splice

ThPatQuote
A Template Haskell pattern quotation [p| (a,b) |]

PatSyn
A pattern synonym declaration

isPatSynCtxt :: HsMatchContext id -> Bool

data HsStmtContext id

Haskell Statement Context. It expects to be parameterised with one of RdrName, Name or Id

Constructors

ListComp

MonadComp

DoExpr
do { ... }

MDoExpr
mdo { ... } ie recursive do-expression

ArrowExpr
do-notation in an arrow-command context

GhciStmtCtxt
A command-line Stmt in GHCi pat <- rhs

PatGuard (HsMatchContext id)
Pattern guard for specified thing

ParStmtCtxt (HsStmtContext id)
A branch of a parallel stmt

TransStmtCtxt (HsStmtContext id)
A branch of a transform stmt

isListCompExpr :: HsStmtContext id -> Bool

isMonadFailStmtContext :: HsStmtContext id -> Bool

Should pattern match failure in a HsStmtContext be desugared using MonadFail?

matchSeparator :: HsMatchContext id -> SDoc

pprMatchContext :: (Outputable (NameOrRdrName id), Outputable id) => HsMatchContext id -> SDoc

pprMatchContextNoun :: (Outputable (NameOrRdrName id), Outputable id) => HsMatchContext id -> SDoc

pprAStmtContext :: (Outputable id, Outputable (NameOrRdrName id)) => HsStmtContext id -> SDoc

pprStmtContext :: (Outputable id, Outputable (NameOrRdrName id)) => HsStmtContext id -> SDoc

matchContextErrString :: Outputable id => HsMatchContext id -> SDoc

pprMatchInCtxt :: (OutputableBndrId (GhcPass idR), Outputable (NameOrRdrName (NameOrRdrName (IdP (GhcPass idR)))), Outputable body) => Match (GhcPass idR) body -> SDoc

pprStmtInCtxt :: (OutputableBndrId (GhcPass idL), OutputableBndrId (GhcPass idR), Outputable body) => HsStmtContext (IdP (GhcPass idL)) -> StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc