HsImpExp

Documentation

type LImportDecl pass

Arguments

= Located (ImportDecl pass)
When in a list this may have

AnnKeywordId : AnnSemi

Located Import Declaration

data ImportDecl pass

Import Declaration

A single Haskell import declaration.

Constructors

ImportDecl

Fields

ideclExt :: XCImportDecl pass

ideclSourceSrc :: SourceText

ideclName :: Located ModuleName

Module name.

ideclPkgQual :: Maybe StringLiteral

Package qualifier.

ideclSource :: Bool

True = {-# SOURCE #-} import

ideclSafe :: Bool

True => safe import

ideclQualified :: Bool

True => qualified

ideclImplicit :: Bool

True => implicit import (of Prelude)

ideclAs :: Maybe (Located ModuleName)

as Module

ideclHiding :: Maybe (Bool, Located [LIE pass])

(True => hiding, names)

XImportDecl (XXImportDecl pass)
AnnKeywordIds

AnnImport

AnnOpen, AnnClose for ideclSource

AnnSafe,AnnQualified, AnnPackageName,AnnAs, AnnVal

AnnHiding,AnnOpen, AnnClose attached to location in ideclHiding

simpleImportDecl :: ModuleName -> ImportDecl (GhcPass p)

data IEWrappedName name

A name in an import or export specification which may have adornments. Used primarily for accurate pretty printing of ParsedSource, and API Annotation placement.

Constructors

IEName (Located name)
no extra

IEPattern (Located name)
pattern X

IEType (Located name)
type (:+:)

type LIEWrappedName name = Located (IEWrappedName name)

Located name with possible adornment - AnnKeywordIds : AnnType, AnnPattern

type LIE pass

Arguments

= Located (IE pass)
When in a list this may have

AnnKeywordId : AnnComma

Located Import or Export

data IE pass

Imported or exported entity.

Constructors

IEVar (XIEVar pass) (LIEWrappedName (IdP pass))
Imported or Exported Variable

IEThingAbs (XIEThingAbs pass) (LIEWrappedName (IdP pass))
Imported or exported Thing with Absent list

The thing is a Class/Type (can't tell) - AnnKeywordIds : AnnPattern, AnnType,AnnVal

IEThingAll (XIEThingAll pass) (LIEWrappedName (IdP pass))
Imported or exported Thing with All imported or exported

The thing is a ClassType and the All refers to methodsconstructors

AnnKeywordIds : AnnOpen, AnnDotdot,AnnClose, AnnType

IEThingWith (XIEThingWith pass) (LIEWrappedName (IdP pass)) IEWildcard [LIEWrappedName (IdP pass)] [Located (FieldLbl (IdP pass))]
Imported or exported Thing With given imported or exported

The thing is a Class/Type and the imported or exported things are methods/constructors and record fields; see Note [IEThingWith] - AnnKeywordIds : AnnOpen, AnnClose, AnnComma, AnnType

IEModuleContents (XIEModuleContents pass) (Located ModuleName)
Imported or exported module contents

(Export Only)

AnnKeywordIds : AnnModule

IEGroup (XIEGroup pass) Int HsDocString
Doc section heading

IEDoc (XIEDoc pass) HsDocString
Some documentation

IEDocNamed (XIEDocNamed pass) String
Reference to named doc

XIE (XXIE pass)

data IEWildcard

Imported or Exported Wildcard

Constructors

NoIEWildcard

IEWildcard Int

ieName :: IE pass -> IdP pass

ieNames :: IE pass -> [IdP pass]

ieWrappedName :: IEWrappedName name -> name

ieLWrappedName :: LIEWrappedName name -> Located name

replaceWrappedName :: IEWrappedName name1 -> name2 -> IEWrappedName name2

replaceLWrappedName :: LIEWrappedName name1 -> name2 -> LIEWrappedName name2

pprImpExp :: (HasOccName name, OutputableBndr name) => name -> SDoc