HsLit

Documentation

data HsLit x

Haskell Literal

Constructors

HsChar (XHsChar x) Char
Character

HsCharPrim (XHsCharPrim x) Char
Unboxed character

HsString (XHsString x) FastString
String

HsStringPrim (XHsStringPrim x) ByteString
Packed bytes

HsInt (XHsInt x) IntegralLit
Genuinely an Int; arises from TcGenDeriv, and from TRANSLATION

HsIntPrim (XHsIntPrim x) Integer
literal Int#

HsWordPrim (XHsWordPrim x) Integer
literal Word#

HsInt64Prim (XHsInt64Prim x) Integer
literal Int64#

HsWord64Prim (XHsWord64Prim x) Integer
literal Word64#

HsInteger (XHsInteger x) Integer Type
Genuinely an integer; arises only from TRANSLATION (overloaded literals are done with HsOverLit)

HsRat (XHsRat x) FractionalLit Type
Genuinely a rational; arises only from TRANSLATION (overloaded literals are done with HsOverLit)

HsFloatPrim (XHsFloatPrim x) FractionalLit
Unboxed Float

HsDoublePrim (XHsDoublePrim x) FractionalLit
Unboxed Double

XLit (XXLit x)

data HsOverLit p

Haskell Overloaded Literal

Constructors

OverLit

Fields

ol_ext :: XOverLit p

ol_val :: OverLitVal

ol_witness :: HsExpr p

XOverLit (XXOverLit p)

data OverLitTc

Constructors

OverLitTc

Fields

ol_rebindable :: Bool

ol_type :: Type

data OverLitVal

Overloaded Literal Value

Constructors

HsIntegral !IntegralLit
Integer-looking literals;

HsFractional !FractionalLit
Frac-looking literals

HsIsString !SourceText !FastString
String-looking literals

negateOverLitVal :: OverLitVal -> OverLitVal

overLitType :: HsOverLit GhcTc -> Type

convertLit :: ConvertIdX a b => HsLit a -> HsLit b

Convert a literal from one index type to another, updating the annotations according to the relevant Convertable instance

pp_st_suffix :: SourceText -> SDoc -> SDoc -> SDoc

pmPprHsLit :: HsLit (GhcPass x) -> SDoc

pmPprHsLit pretty prints literals and is used when pretty printing pattern match warnings. All are printed the same (i.e., without hashes if they are primitive and not wrapped in constructors if they are boxed). This happens mainly for too reasons: * We do not want to expose their internal representation * The warnings become too messy

hsLitNeedsParens :: PprPrec -> HsLit x -> Bool

hsLitNeedsParens p l returns True if a literal l needs to be parenthesized under precedence p.

hsOverLitNeedsParens :: PprPrec -> HsOverLit x -> Bool

hsOverLitNeedsParens p ol returns True if an overloaded literal ol needs to be parenthesized under precedence p.