HsPat

Documentation

data Pat p

Pattern

AnnKeywordId : AnnBang

Constructors

WildPat (XWildPat p)
Wildcard Pattern The sole reason for a type on a WildPat is to support hsPatType :: Pat Id -> Type

VarPat (XVarPat p) (Located (IdP p))
Variable Pattern

LazyPat (XLazyPat p) (LPat p)
Lazy Pattern ^ - AnnKeywordId : AnnTilde

AsPat (XAsPat p) (Located (IdP p)) (LPat p)
As pattern ^ - AnnKeywordId : AnnAt

ParPat (XParPat p) (LPat p)
Parenthesised pattern See Note [Parens in HsSyn] in HsExpr ^ - AnnKeywordId : AnnOpen '(', AnnClose ')'

BangPat (XBangPat p) (LPat p)
Bang pattern ^ - AnnKeywordId : AnnBang

ListPat (XListPat p) [LPat p]
Syntactic List

AnnKeywordId : AnnOpen '[', AnnClose ']'

TuplePat (XTuplePat p) [LPat p] Boxity
Tuple sub-patterns

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

SumPat (XSumPat p) (LPat p) ConTag Arity
Anonymous sum pattern

AnnKeywordId : AnnOpen '(#', AnnClose '#)'

ConPatIn (Located (IdP p)) (HsConPatDetails p)
Constructor Pattern In

ConPatOut
Constructor Pattern Out

Fields

pat_con :: Located ConLike

pat_arg_tys :: [Type]

pat_tvs :: [TyVar]

pat_dicts :: [EvVar]

pat_binds :: TcEvBinds

pat_args :: HsConPatDetails p

pat_wrap :: HsWrapper

ViewPat (XViewPat p) (LHsExpr p) (LPat p)
View Pattern

SplicePat (XSplicePat p) (HsSplice p)
Splice Pattern (Includes quasi-quotes)

LitPat (XLitPat p) (HsLit p)
Literal Pattern Used for *non-overloaded* literal patterns: Int, Int, Char, String, etc.

NPat (XNPat p) (Located (HsOverLit p)) (Maybe (SyntaxExpr p)) (SyntaxExpr p)
Natural Pattern

AnnKeywordId : AnnVal +

NPlusKPat (XNPlusKPat p) (Located (IdP p)) (Located (HsOverLit p)) (HsOverLit p) (SyntaxExpr p) (SyntaxExpr p)
n+k pattern

SigPat (XSigPat p) (LPat p)
Pattern with a type signature

CoPat (XCoPat p) HsWrapper (Pat p) Type
Coercion Pattern

XPat (XXPat p)
Trees that Grow extension point for new constructors

type InPat p = LPat p

type OutPat p = LPat p

type LPat p = Located (Pat p)

data ListPatTc

Constructors

ListPatTc Type (Maybe (Type, SyntaxExpr GhcTc))

type HsConPatDetails p = HsConDetails (LPat p) (HsRecFields p (LPat p))

Haskell Constructor Pattern Details

hsConPatArgs :: HsConPatDetails p -> [LPat p]

data HsRecFields p arg

Haskell Record Fields

HsRecFields is used only for patterns and expressions (not data type declarations)

Constructors

HsRecFields

Fields

rec_flds :: [LHsRecField p arg]

rec_dotdot :: Maybe Int

data HsRecField' id arg

Haskell Record Field

AnnKeywordId : AnnEqual,

For details on above see note [Api annotations] in ApiAnnotation

Constructors

HsRecField

Fields

hsRecFieldLbl :: Located id

hsRecFieldArg :: arg

Filled in by renamer when punning

hsRecPun :: Bool

Note [Punning]

type LHsRecField' p arg = Located (HsRecField' p arg)

Located Haskell Record Field

type HsRecField p arg = HsRecField' (FieldOcc p) arg

Haskell Record Field

type LHsRecField p arg = Located (HsRecField p arg)

Located Haskell Record Field

type HsRecUpdField p = HsRecField' (AmbiguousFieldOcc p) (LHsExpr p)

Haskell Record Update Field

type LHsRecUpdField p = Located (HsRecUpdField p)

Located Haskell Record Update Field

hsRecFields :: HsRecFields p arg -> [XCFieldOcc p]

hsRecFieldSel :: HsRecField pass arg -> Located (XCFieldOcc pass)

hsRecFieldId :: HsRecField GhcTc arg -> Located Id

hsRecFieldsArgs :: HsRecFields p arg -> [arg]

hsRecUpdFieldId :: HsRecField' (AmbiguousFieldOcc GhcTc) arg -> Located Id

hsRecUpdFieldOcc :: HsRecField' (AmbiguousFieldOcc GhcTc) arg -> LFieldOcc GhcTc

hsRecUpdFieldRdr :: HsRecUpdField (GhcPass p) -> Located RdrName

mkPrefixConPat :: DataCon -> [OutPat p] -> [Type] -> OutPat p

mkCharLitPat :: SourceText -> Char -> OutPat (GhcPass p)

mkNilPat :: Type -> OutPat p

looksLazyPatBind :: HsBind p -> Bool

isBangedLPat :: LPat p -> Bool

patNeedsParens :: PprPrec -> Pat p -> Bool

patNeedsParens p pat returns True if the pattern pat needs parentheses under precedence p.

parenthesizePat :: PprPrec -> LPat (GhcPass p) -> LPat (GhcPass p)

parenthesizePat p pat checks if patNeedsParens p pat is true, and if so, surrounds pat with a ParPat. Otherwise, it simply returns pat.

isIrrefutableHsPat :: OutputableBndrId p => LPat p -> Bool

collectEvVarsPat :: Pat GhcTc -> Bag EvVar

collectEvVarsPats :: [Pat GhcTc] -> Bag EvVar

pprParendLPat :: OutputableBndrId (GhcPass p) => PprPrec -> LPat (GhcPass p) -> SDoc

pprConArgs :: OutputableBndrId (GhcPass p) => HsConPatDetails (GhcPass p) -> SDoc