HsTypes

Documentation

data HsType pass

Haskell Type

Constructors

HsForAllTy

AnnKeywordId : AnnForall, AnnDot,AnnDarrow For details on above see note [Api annotations] in ApiAnnotation

Fields

hst_xforall :: XForAllTy pass

hst_bndrs :: [LHsTyVarBndr pass]

hst_body :: LHsType pass

HsQualTy

Fields

hst_xqual :: XQualTy pass

hst_ctxt :: LHsContext pass

hst_body :: LHsType pass

HsTyVar (XTyVar pass) Promoted (Located (IdP pass))

AnnKeywordId : None

HsAppTy (XAppTy pass) (LHsType pass) (LHsType pass)

AnnKeywordId : None

HsFunTy (XFunTy pass) (LHsType pass) (LHsType pass)

AnnKeywordId : AnnRarrow,

HsListTy (XListTy pass) (LHsType pass)

AnnKeywordId : AnnOpen '[', AnnClose ']'

HsTupleTy (XTupleTy pass) HsTupleSort [LHsType pass]

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

HsSumTy (XSumTy pass) [LHsType pass]

AnnKeywordId : AnnOpen '(#', AnnClose '#)'@

HsOpTy (XOpTy pass) (LHsType pass) (Located (IdP pass)) (LHsType pass)

AnnKeywordId : None

HsParTy (XParTy pass) (LHsType pass)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsIParamTy (XIParamTy pass) (Located HsIPName) (LHsType pass)
(?x :: ty)

AnnKeywordId : AnnDcolon

HsStarTy (XStarTy pass) Bool

AnnKeywordId : None

HsKindSig (XKindSig pass) (LHsType pass) (LHsKind pass)
(ty :: kind)

AnnKeywordId : AnnOpen '(', AnnDcolon,AnnClose ')'

HsSpliceTy (XSpliceTy pass) (HsSplice pass)

AnnKeywordId : AnnOpen '$(', AnnClose ')'

HsDocTy (XDocTy pass) (LHsType pass) LHsDocString

AnnKeywordId : None

HsBangTy (XBangTy pass) HsSrcBang (LHsType pass)

AnnKeywordId : AnnOpen '{-# UNPACK' or '{-# NOUNPACK', AnnClose '#-}' AnnBang '!'

HsRecTy (XRecTy pass) [LConDeclField pass]

AnnKeywordId : AnnOpen '{', AnnClose '}'

HsExplicitListTy (XExplicitListTy pass) Promoted [LHsType pass]

AnnKeywordId : AnnOpen "'[", AnnClose ']'

HsExplicitTupleTy (XExplicitTupleTy pass) [LHsType pass]

AnnKeywordId : AnnOpen "'(", AnnClose ')'

HsTyLit (XTyLit pass) HsTyLit

AnnKeywordId : None

HsWildCardTy (XWildCardTy pass)

AnnKeywordId : None

XHsType (XXType pass)

data NewHsTypeX

Constructors

NHsCoreTy Type

type LHsType pass

Arguments

= Located (HsType pass)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Type

type HsKind pass = HsType pass

Haskell Kind

type LHsKind pass

Arguments

= Located (HsKind pass)
AnnKeywordId : AnnDcolon

Located Haskell Kind

data HsTyVarBndr pass

Haskell Type Variable Binder

Constructors

UserTyVar (XUserTyVar pass) (Located (IdP pass))

KindedTyVar (XKindedTyVar pass) (Located (IdP pass)) (LHsKind pass)

AnnKeywordId : AnnOpen, AnnDcolon, AnnClose

XTyVarBndr (XXTyVarBndr pass)

type LHsTyVarBndr pass = Located (HsTyVarBndr pass)

Located Haskell Type Variable Binder

data LHsQTyVars pass

Located Haskell Quantified Type Variables

Constructors

HsQTvs

Fields

hsq_ext :: XHsQTvs pass

hsq_explicit :: [LHsTyVarBndr pass]

XLHsQTyVars (XXLHsQTyVars pass)

data HsQTvsRn

Constructors

HsQTvsRn

Fields

hsq_implicit :: [Name]

hsq_dependent :: NameSet

data HsImplicitBndrs pass thing

Haskell Implicit Binders

Constructors

HsIB

Fields

hsib_ext :: XHsIB pass thing

hsib_body :: thing

XHsImplicitBndrs (XXHsImplicitBndrs pass thing)

data HsIBRn

Constructors

HsIBRn

Fields

hsib_vars :: [Name]

hsib_closed :: Bool

data HsWildCardBndrs pass thing

Haskell Wildcard Binders

Constructors

HsWC

Fields

hswc_ext :: XHsWC pass thing

hswc_body :: thing

XHsWildCardBndrs (XXHsWildCardBndrs pass thing)

type LHsSigType pass = HsImplicitBndrs pass (LHsType pass)

Located Haskell Signature Type

type LHsSigWcType pass = HsWildCardBndrs pass (LHsSigType pass)

Located Haskell Signature Wildcard Type

type LHsWcType pass = HsWildCardBndrs pass (LHsType pass)

Located Haskell Wildcard Type

data HsTupleSort

Haskell Tuple Sort

Constructors

HsUnboxedTuple

HsBoxedTuple

HsConstraintTuple

HsBoxedOrConstraintTuple

data Promoted

Promoted data types.

Constructors

Promoted

NotPromoted

type HsContext pass = [LHsType pass]

Haskell Context

type LHsContext pass

Arguments

= Located (HsContext pass)
AnnKeywordId : AnnUnit

Located Haskell Context

data HsTyLit

Haskell Type Literal

Constructors

HsNumTy SourceText Integer

HsStrTy SourceText FastString

newtype HsIPName

These names are used early on to store the names of implicit parameters. They completely disappear after type-checking.

Constructors

HsIPName FastString

hsIPNameFS :: HsIPName -> FastString

type LBangType pass = Located (BangType pass)

Located Bang Type

type BangType pass = HsType pass

Bang Type

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as the user wrote them in the source code.

(HsSrcBang _ SrcUnpack SrcLazy) and (HsSrcBang _ SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsSrcBang _ NoSrcUnpack SrcLazy)

Constructors

HsSrcBang SourceText SrcUnpackedness SrcStrictness

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field, or one with an unlifted type

HsStrict
Strict but not unpacked field

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie '~'

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{--} specified

SrcNoUnpack
{--} specified

NoSrcUnpack
no unpack pragma

getBangType :: LHsType a -> LHsType a

getBangStrictness :: LHsType a -> HsSrcBang

data ConDeclField pass

Constructor Declaration Field

Constructors

ConDeclField

AnnKeywordId : AnnDcolon

Fields

cd_fld_ext :: XConDeclField pass

cd_fld_names :: [LFieldOcc pass]

See Note [ConDeclField passs]

cd_fld_type :: LBangType pass

cd_fld_doc :: Maybe LHsDocString

XConDeclField (XXConDeclField pass)

type LConDeclField pass

Arguments

= Located (ConDeclField pass)
May have AnnKeywordId : AnnComma when in a list

Located Constructor Declaration Field

pprConDeclFields :: OutputableBndrId (GhcPass p) => [LConDeclField (GhcPass p)] -> SDoc

data HsConDetails arg rec

Haskell Constructor Details

Constructors

PrefixCon [arg]

RecCon rec

InfixCon arg arg

data FieldOcc pass

Field Occurrence

Represents an *occurrence* of an unambiguous field. We store both the RdrName the user originally wrote, and after the renamer, the selector function.

Constructors

FieldOcc

Fields

extFieldOcc :: XCFieldOcc pass

rdrNameFieldOcc :: Located RdrName

See Note [Located RdrNames] in HsExpr

XFieldOcc (XXFieldOcc pass)

type LFieldOcc pass = Located (FieldOcc pass)

Located Field Occurrence

mkFieldOcc :: Located RdrName -> FieldOcc GhcPs

data AmbiguousFieldOcc pass

Ambiguous Field Occurrence

Represents an *occurrence* of a field that is potentially ambiguous after the renamer, with the ambiguity resolved by the typechecker. We always store the RdrName that the user originally wrote, and store the selector function after the renamer (for unambiguous occurrences) or the typechecker (for ambiguous occurrences).

See Note [HsRecField and HsRecUpdField] in HsPat and Note [Disambiguating record fields] in TcExpr. See Note [Located RdrNames] in HsExpr

Constructors

Unambiguous (XUnambiguous pass) (Located RdrName)

Ambiguous (XAmbiguous pass) (Located RdrName)

XAmbiguousFieldOcc (XXAmbiguousFieldOcc pass)

mkAmbiguousFieldOcc :: Located RdrName -> AmbiguousFieldOcc GhcPs

rdrNameAmbiguousFieldOcc :: AmbiguousFieldOcc (GhcPass p) -> RdrName

selectorAmbiguousFieldOcc :: AmbiguousFieldOcc GhcTc -> Id

unambiguousFieldOcc :: AmbiguousFieldOcc GhcTc -> FieldOcc GhcTc

ambiguousFieldOcc :: FieldOcc GhcTc -> AmbiguousFieldOcc GhcTc

newtype HsWildCardInfo

Constructors

AnonWildCard (Located Name)

mkAnonWildCardTy :: HsType GhcPs

pprAnonWildCard :: SDoc

wildCardName :: HsWildCardInfo -> Name

sameWildCard :: Located HsWildCardInfo -> Located HsWildCardInfo -> Bool

mkHsImplicitBndrs :: thing -> HsImplicitBndrs GhcPs thing

mkHsWildCardBndrs :: thing -> HsWildCardBndrs GhcPs thing

hsImplicitBody :: HsImplicitBndrs pass thing -> thing

mkEmptyImplicitBndrs :: thing -> HsImplicitBndrs GhcRn thing

mkEmptyWildCardBndrs :: thing -> HsWildCardBndrs GhcRn thing

mkHsQTvs :: [LHsTyVarBndr GhcPs] -> LHsQTyVars GhcPs

hsQTvExplicit :: LHsQTyVars pass -> [LHsTyVarBndr pass]

emptyLHsQTvs :: LHsQTyVars GhcRn

isEmptyLHsQTvs :: LHsQTyVars GhcRn -> Bool

isHsKindedTyVar :: HsTyVarBndr pass -> Bool

Does this HsTyVarBndr come with an explicit kind annotation?

hsTvbAllKinded :: LHsQTyVars pass -> Bool

Do all type variables in this LHsQTyVars come with kind annotations?

isLHsForAllTy :: LHsType p -> Bool

hsScopedTvs :: LHsSigType GhcRn -> [Name]

hsWcScopedTvs :: LHsSigWcType GhcRn -> [Name]

dropWildCards :: LHsSigWcType pass -> LHsSigType pass

hsTyVarName :: HsTyVarBndr pass -> IdP pass

hsAllLTyVarNames :: LHsQTyVars GhcRn -> [Name]

hsLTyVarLocNames :: LHsQTyVars pass -> [Located (IdP pass)]

hsLTyVarName :: LHsTyVarBndr pass -> IdP pass

hsLTyVarLocName :: LHsTyVarBndr pass -> Located (IdP pass)

hsExplicitLTyVarNames :: LHsQTyVars pass -> [IdP pass]

splitLHsInstDeclTy :: LHsSigType GhcRn -> ([Name], LHsContext GhcRn, LHsType GhcRn)

getLHsInstDeclHead :: LHsSigType pass -> LHsType pass

getLHsInstDeclClass_maybe :: LHsSigType (GhcPass p) -> Maybe (Located (IdP (GhcPass p)))

splitLHsPatSynTy :: LHsType pass -> ([LHsTyVarBndr pass], LHsContext pass, [LHsTyVarBndr pass], LHsContext pass, LHsType pass)

splitLHsForAllTy :: LHsType pass -> ([LHsTyVarBndr pass], LHsType pass)

splitLHsQualTy :: LHsType pass -> (LHsContext pass, LHsType pass)

splitLHsSigmaTy :: LHsType pass -> ([LHsTyVarBndr pass], LHsContext pass, LHsType pass)

splitHsFunType :: LHsType GhcRn -> ([LHsType GhcRn], LHsType GhcRn)

splitHsAppTys :: LHsType GhcRn -> [LHsType GhcRn] -> (LHsType GhcRn, [LHsType GhcRn])

hsTyGetAppHead_maybe :: LHsType (GhcPass p) -> Maybe (Located (IdP (GhcPass p)), [LHsType (GhcPass p)])

mkHsOpTy :: LHsType (GhcPass p) -> Located (IdP (GhcPass p)) -> LHsType (GhcPass p) -> HsType (GhcPass p)

mkHsAppTy :: LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

mkHsAppTys :: LHsType (GhcPass p) -> [LHsType (GhcPass p)] -> LHsType (GhcPass p)

ignoreParens :: LHsType pass -> LHsType pass

hsSigType :: LHsSigType pass -> LHsType pass

hsSigWcType :: LHsSigWcType pass -> LHsType pass

hsLTyVarBndrToType :: LHsTyVarBndr (GhcPass p) -> LHsType (GhcPass p)

Convert a LHsTyVarBndr to an equivalent LHsType.

hsLTyVarBndrsToTypes :: LHsQTyVars (GhcPass p) -> [LHsType (GhcPass p)]

Convert a LHsTyVarBndrs to a list of types. Works on *type* variable only, no kind vars.

pprHsType :: OutputableBndrId (GhcPass p) => HsType (GhcPass p) -> SDoc

pprHsForAll :: OutputableBndrId (GhcPass p) => [LHsTyVarBndr (GhcPass p)] -> LHsContext (GhcPass p) -> SDoc

pprHsForAllTvs :: OutputableBndrId (GhcPass p) => [LHsTyVarBndr (GhcPass p)] -> SDoc

pprHsForAllExtra :: OutputableBndrId (GhcPass p) => Maybe SrcSpan -> [LHsTyVarBndr (GhcPass p)] -> LHsContext (GhcPass p) -> SDoc

Version of pprHsForAll that can also print an extra-constraints wildcard, e.g. _ => a -> Bool or (Show a, _) => a -> String. This underscore will be printed when the 'Maybe SrcSpan' argument is a Just containing the location of the extra-constraints wildcard. A special function for this is needed, as the extra-constraints wildcard is removed from the actual context and type, and stored in a separate field, thus just printing the type will not print the extra-constraints wildcard.

pprHsContext :: OutputableBndrId (GhcPass p) => HsContext (GhcPass p) -> SDoc

pprHsContextNoArrow :: OutputableBndrId (GhcPass p) => HsContext (GhcPass p) -> SDoc

pprHsContextMaybe :: OutputableBndrId (GhcPass p) => HsContext (GhcPass p) -> Maybe SDoc

hsTypeNeedsParens :: PprPrec -> HsType pass -> Bool

hsTypeNeedsParens p t returns True if the type t needs parentheses under precedence p.

parenthesizeHsType :: PprPrec -> LHsType (GhcPass p) -> LHsType (GhcPass p)

parenthesizeHsType p ty checks if hsTypeNeedsParens p ty is true, and if so, surrounds ty with an HsParTy. Otherwise, it simply returns ty.