Id

Description

GHC uses several kinds of name internally:

OccName: see OccName

RdrName: see RdrName

Name: see Name

Id represents names that not only have a Name but also a Type and some additional details (a IdInfo and one of LocalIdDetails or GlobalIdDetails) that are added, modified and inspected by various compiler passes. These Var names may either be global or local, see Var

Var: see Var

The main types

data Var

Variable

Essentially a typed Name, that may also contain some additional information about the Var and it's use sites.

type Id = Var

Identifier

isId :: Var -> Bool

In and Out variants

type InVar = Var

type InId = Id

type OutVar = Var

type OutId = Id

Simple construction

mkGlobalId :: IdDetails -> Name -> Type -> IdInfo -> Id

For an explanation of global vs. local Ids, see Var

mkVanillaGlobal :: Name -> Type -> Id

Make a global Id without any extra information at all

mkVanillaGlobalWithInfo :: Name -> Type -> IdInfo -> Id

Make a global Id with no global information but some generic IdInfo

mkLocalId :: Name -> Type -> Id

For an explanation of global vs. local Ids, see Var

mkLocalCoVar :: Name -> Type -> CoVar

Make a local CoVar

mkLocalIdOrCoVar :: Name -> Type -> Id

Like mkLocalId, but checks the type to see if it should make a covar

mkLocalIdOrCoVarWithInfo :: Name -> Type -> IdInfo -> Id

Make a local id, with the IdDetails set to CoVarId if the type indicates so.

mkLocalIdWithInfo :: Name -> Type -> IdInfo -> Id

mkExportedLocalId :: IdDetails -> Name -> Type -> Id

Create a local Id that is marked as exported. This prevents things attached to it from being removed as dead code. See Note [Exported LocalIds]

mkExportedVanillaId :: Name -> Type -> Id

mkSysLocal :: FastString -> Unique -> Type -> Id

Create a system local Id. These are local Ids (see Var) that are created by the compiler out of thin air

mkSysLocalM :: MonadUnique m => FastString -> Type -> m Id

mkSysLocalOrCoVar :: FastString -> Unique -> Type -> Id

Like mkSysLocal, but checks to see if we have a covar type

mkSysLocalOrCoVarM :: MonadUnique m => FastString -> Type -> m Id

mkUserLocal :: OccName -> Unique -> Type -> SrcSpan -> Id

Create a user local Id. These are local Ids (see Var) with a name and location that the user might recognize

mkUserLocalOrCoVar :: OccName -> Unique -> Type -> SrcSpan -> Id

Like mkUserLocal, but checks if we have a coercion type

mkTemplateLocals :: [Type] -> [Id]

Create a template local for a series of types

mkTemplateLocalsNum :: Int -> [Type] -> [Id]

Create a template local for a series of type, but start from a specified template local

mkTemplateLocal :: Int -> Type -> Id

Create a template local: a family of system local Ids in bijection with Ints, typically used in unfoldings

mkWorkerId :: Unique -> Id -> Type -> Id

Workers get local names. CoreTidy will externalise these if necessary

Taking an Id apart

idName :: Id -> Name

idType :: Id -> Kind

idUnique :: Id -> Unique

idInfo :: HasDebugCallStack => Id -> IdInfo

idDetails :: Id -> IdDetails

recordSelectorTyCon :: Id -> RecSelParent

If the Id is that for a record selector, extract the sel_tycon. Panic otherwise.

Modifying an Id

setIdName :: Id -> Name -> Id

setIdUnique :: Id -> Unique -> Id

setIdType :: Id -> Type -> Id

Not only does this set the Id Type, it also evaluates the type to try and reduce space usage

setIdExported :: Id -> Id

setIdNotExported :: Id -> Id

globaliseId :: Id -> Id

If it's a local, make it global

localiseId :: Id -> Id

setIdInfo :: Id -> IdInfo -> Id

lazySetIdInfo :: Id -> IdInfo -> Id

modifyIdInfo :: (IdInfo -> IdInfo) -> Id -> Id

maybeModifyIdInfo :: Maybe IdInfo -> Id -> Id

zapLamIdInfo :: Id -> Id

zapIdDemandInfo :: Id -> Id

zapIdUsageInfo :: Id -> Id

zapIdUsageEnvInfo :: Id -> Id

zapIdUsedOnceInfo :: Id -> Id

zapIdTailCallInfo :: Id -> Id

zapFragileIdInfo :: Id -> Id

zapIdStrictness :: Id -> Id

transferPolyIdInfo :: Id -> [Var] -> Id -> Id

Predicates on Ids

isImplicitId :: Id -> Bool

isImplicitId tells whether an Ids info is implied by other declarations, so we don't need to put its signature in an interface file, even if it's mentioned in some other interface unfolding.

isDeadBinder :: Id -> Bool

isStrictId :: Id -> Bool

This predicate says whether the Id has a strict demand placed on it or has a type such that it can always be evaluated strictly (i.e an unlifted type, as of GHC 7.6). We need to check separately whether the Id has a so-called "strict type" because if the demand for the given id hasn't been computed yet but id has a strict type, we still want isStrictId id to be True.

isExportedId :: Var -> Bool

isExportedIdVar means "don't throw this away"

isLocalId :: Var -> Bool

isGlobalId :: Var -> Bool

isRecordSelector :: Id -> Bool

isNaughtyRecordSelector :: Id -> Bool

isPatSynRecordSelector :: Id -> Bool

isDataConRecordSelector :: Id -> Bool

isClassOpId_maybe :: Id -> Maybe Class

isDFunId :: Id -> Bool

isPrimOpId :: Id -> Bool

isPrimOpId_maybe :: Id -> Maybe PrimOp

isFCallId :: Id -> Bool

isFCallId_maybe :: Id -> Maybe ForeignCall

isDataConWorkId :: Id -> Bool

isDataConWorkId_maybe :: Id -> Maybe DataCon

isDataConId_maybe :: Id -> Maybe DataCon

idDataCon :: Id -> DataCon

Get from either the worker or the wrapper Id to the DataCon. Currently used only in the desugarer.

INVARIANT: idDataCon (dataConWrapId d) = d: remember, dataConWrapId can return either the wrapper or the worker

isConLikeId :: Id -> Bool

isBottomingId :: Var -> Bool

Returns true if an application to n args would diverge

idIsFrom :: Module -> Id -> Bool

hasNoBinding :: Id -> Bool

Returns True of an Id which may not have a binding, even though it is defined in this module.

Evidence variables

type DictId = EvId

Dictionary Identifier

isDictId :: Id -> Bool

isEvVar :: Var -> Bool

Join variables

type JoinId = Id

isJoinId :: Var -> Bool

isJoinId_maybe :: Var -> Maybe JoinArity

idJoinArity :: JoinId -> JoinArity

asJoinId :: Id -> JoinArity -> JoinId infixl 1

asJoinId_maybe :: Id -> Maybe JoinArity -> Id infixl 1

zapJoinId :: Id -> Id

Inline pragma stuff

idInlinePragma :: Id -> InlinePragma

setInlinePragma :: Id -> InlinePragma -> Id infixl 1

modifyInlinePragma :: Id -> (InlinePragma -> InlinePragma) -> Id

idInlineActivation :: Id -> Activation

setInlineActivation :: Id -> Activation -> Id infixl 1

idRuleMatchInfo :: Id -> RuleMatchInfo

One-shot lambdas

isOneShotBndr :: Var -> Bool

Returns whether the lambda associated with the Id is certainly applied at most once This one is the "business end", called externally. It works on type variables as well as Ids, returning True Its main purpose is to encapsulate the Horrible State Hack See Note [The state-transformer hack] in CoreArity

isProbablyOneShotLambda :: Id -> Bool

setOneShotLambda :: Id -> Id

clearOneShotLambda :: Id -> Id

updOneShotInfo :: Id -> OneShotInfo -> Id

setIdOneShotInfo :: Id -> OneShotInfo -> Id infixl 1

isStateHackType :: Type -> Bool

stateHackOneShot :: OneShotInfo

Should we apply the state hack to values of this Type?

typeOneShot :: Type -> OneShotInfo

Reading IdInfo fields

idArity :: Id -> Arity

idCallArity :: Id -> Arity

idFunRepArity :: Id -> RepArity

idUnfolding :: Id -> Unfolding

realIdUnfolding :: Id -> Unfolding

idSpecialisation :: Id -> RuleInfo

idCoreRules :: Id -> [CoreRule]

idHasRules :: Id -> Bool

idCafInfo :: Id -> CafInfo infixl 1

idOneShotInfo :: Id -> OneShotInfo

idStateHackOneShotInfo :: Id -> OneShotInfo

Like idOneShotInfo, but taking the Horrible State Hack in to account See Note [The state-transformer hack] in CoreArity

idOccInfo :: Id -> OccInfo

isNeverLevPolyId :: Id -> Bool

Writing IdInfo fields

setIdUnfolding :: Id -> Unfolding -> Id infixl 1

setCaseBndrEvald :: StrictnessMark -> Id -> Id

setIdArity :: Id -> Arity -> Id infixl 1

setIdCallArity :: Id -> Arity -> Id infixl 1

setIdSpecialisation :: Id -> RuleInfo -> Id infixl 1

setIdCafInfo :: Id -> CafInfo -> Id

setIdOccInfo :: Id -> OccInfo -> Id infixl 1

zapIdOccInfo :: Id -> Id

setIdDemandInfo :: Id -> Demand -> Id infixl 1

setIdStrictness :: Id -> StrictSig -> Id infixl 1

idDemandInfo :: Id -> Demand

idStrictness :: Id -> StrictSig