IfaceType

Documentation

type IfExtName = Name

type IfLclName = FastString

data IfaceType

Constructors

IfaceFreeTyVar TyVar

IfaceTyVar IfLclName

IfaceLitTy IfaceTyLit

IfaceAppTy IfaceType IfaceType

IfaceFunTy IfaceType IfaceType

IfaceDFunTy IfaceType IfaceType

IfaceForAllTy IfaceForAllBndr IfaceType

IfaceTyConApp IfaceTyCon IfaceTcArgs

IfaceCastTy IfaceType IfaceCoercion

IfaceCoercionTy IfaceCoercion

IfaceTupleTy TupleSort IsPromoted IfaceTcArgs

type IfacePredType = IfaceType

type IfaceKind = IfaceType

data IfaceCoercion

Constructors

IfaceReflCo Role IfaceType

IfaceFunCo Role IfaceCoercion IfaceCoercion

IfaceTyConAppCo Role IfaceTyCon [IfaceCoercion]

IfaceAppCo IfaceCoercion IfaceCoercion

IfaceForAllCo IfaceTvBndr IfaceCoercion IfaceCoercion

IfaceCoVarCo IfLclName

IfaceAxiomInstCo IfExtName BranchIndex [IfaceCoercion]

IfaceAxiomRuleCo IfLclName [IfaceCoercion]

IfaceUnivCo IfaceUnivCoProv Role IfaceType IfaceType

IfaceSymCo IfaceCoercion

IfaceTransCo IfaceCoercion IfaceCoercion

IfaceNthCo Int IfaceCoercion

IfaceLRCo LeftOrRight IfaceCoercion

IfaceInstCo IfaceCoercion IfaceCoercion

IfaceCoherenceCo IfaceCoercion IfaceCoercion

IfaceKindCo IfaceCoercion

IfaceSubCo IfaceCoercion

IfaceFreeCoVar CoVar

IfaceHoleCo CoVar
See Note [Holes in IfaceCoercion]

data IfaceUnivCoProv

Constructors

IfaceUnsafeCoerceProv

IfacePhantomProv IfaceCoercion

IfaceProofIrrelProv IfaceCoercion

IfacePluginProv String

data IfaceTyCon

Constructors

IfaceTyCon

Fields

ifaceTyConName :: IfExtName

ifaceTyConInfo :: IfaceTyConInfo

data IfaceTyConInfo

Constructors

IfaceTyConInfo

Fields

ifaceTyConIsPromoted :: IsPromoted

ifaceTyConSort :: IfaceTyConSort

data IfaceTyConSort

The various types of TyCons which have special, built-in syntax.

Constructors

IfaceNormalTyCon
a regular tycon

IfaceTupleTyCon !Arity !TupleSort
e.g. (a, b, c) or (). The arity is the tuple width, not the tycon arity (which is twice the width in the case of unboxed tuples).

IfaceSumTyCon !Arity
e.g. (a | b | c)

IfaceEqualityTyCon
A heterogeneous equality TyCon (i.e. eqPrimTyCon, eqReprPrimTyCon, heqTyCon) that is actually being applied to two types of the same kind. This affects pretty-printing only: see Note [Equality predicates in IfaceType]

data IsPromoted

Is a TyCon a promoted data constructor or just a normal type constructor?

Constructors

IsNotPromoted

IsPromoted

data IfaceTyLit

Constructors

IfaceNumTyLit Integer

IfaceStrTyLit FastString

data IfaceTcArgs

Constructors

ITC_Nil

ITC_Vis IfaceType IfaceTcArgs

ITC_Invis IfaceKind IfaceTcArgs

type IfaceContext = [IfacePredType]

data IfaceBndr

Constructors

IfaceIdBndr !IfaceIdBndr

IfaceTvBndr !IfaceTvBndr

data IfaceOneShot

Constructors

IfaceNoOneShot

IfaceOneShot

type IfaceLamBndr = (IfaceBndr, IfaceOneShot)

type IfaceTvBndr = (IfLclName, IfaceKind)

type IfaceIdBndr = (IfLclName, IfaceType)

type IfaceTyConBinder = TyVarBndr IfaceTvBndr TyConBndrVis

type IfaceForAllBndr = TyVarBndr IfaceTvBndr ArgFlag

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [TyVarBndrs, TyVarBinders, TyConBinders, and visibility] in TyCoRep

Constructors

Inferred

Specified

Required

data ShowForAllFlag

Show forall flag

Unconditionally show the forall quantifier with (ShowForAllMust) or when (ShowForAllWhen) the names used are free in the binder or when compiling with -fprint-explicit-foralls.

Constructors

ShowForAllMust

ShowForAllWhen

ifForAllBndrTyVar :: IfaceForAllBndr -> IfaceTvBndr

Extract an IfaceBndr from an IfaceForAllBndr.

ifForAllBndrName :: IfaceForAllBndr -> IfLclName

Extract the variable name from an IfaceForAllBndr.

ifTyConBinderTyVar :: IfaceTyConBinder -> IfaceTvBndr

Extract an IfaceBndr from an IfaceTyConBinder.

ifTyConBinderName :: IfaceTyConBinder -> IfLclName

Extract the variable name from an IfaceTyConBinder.

isIfaceLiftedTypeKind :: IfaceKind -> Bool

tcArgsIfaceTypes :: IfaceTcArgs -> [IfaceType]

pprIfaceType :: IfaceType -> SDoc

pprParendIfaceType :: IfaceType -> SDoc

pprPrecIfaceType :: PprPrec -> IfaceType -> SDoc

pprIfaceContext :: PprPrec -> [IfacePredType] -> SDoc

Prints a context or () if empty You give it the context precedence

pprIfaceContextArr :: [IfacePredType] -> SDoc

Prints "(C a, D b) =>", including the arrow. Used when we want to print a context in a type, so we use funPrec to decide whether to parenthesise a singleton predicate; e.g. Num a => a -> a

pprIfaceIdBndr :: IfaceIdBndr -> SDoc

pprIfaceLamBndr :: IfaceLamBndr -> SDoc

pprIfaceTvBndr :: Bool -> IfaceTvBndr -> SDoc

pprIfaceTyConBinders :: [IfaceTyConBinder] -> SDoc

pprIfaceBndrs :: [IfaceBndr] -> SDoc

pprIfaceTcArgs :: IfaceTcArgs -> SDoc

pprParendIfaceTcArgs :: IfaceTcArgs -> SDoc

pprIfaceForAllPart :: [IfaceForAllBndr] -> [IfacePredType] -> SDoc -> SDoc

pprIfaceForAllPartMust :: [IfaceForAllBndr] -> [IfacePredType] -> SDoc -> SDoc

Like pprIfaceForAllPart, but always uses an explicit forall.

pprIfaceForAll :: [IfaceForAllBndr] -> SDoc

Render the "forall ... ." or "forall ... ->" bit of a type.

pprIfaceSigmaType :: ShowForAllFlag -> IfaceType -> SDoc

pprIfaceTyLit :: IfaceTyLit -> SDoc

pprIfaceCoercion :: IfaceCoercion -> SDoc

pprParendIfaceCoercion :: IfaceCoercion -> SDoc

splitIfaceSigmaTy :: IfaceType -> ([IfaceForAllBndr], [IfacePredType], IfaceType)

pprIfaceTypeApp :: PprPrec -> IfaceTyCon -> IfaceTcArgs -> SDoc

pprUserIfaceForAll :: [IfaceForAllBndr] -> SDoc

pprIfaceCoTcApp :: PprPrec -> IfaceTyCon -> [IfaceCoercion] -> SDoc

pprTyTcApp :: PprPrec -> IfaceTyCon -> IfaceTcArgs -> SDoc

pprIfacePrefixApp :: PprPrec -> SDoc -> [SDoc] -> SDoc

suppressIfaceInvisibles :: DynFlags -> [IfaceTyConBinder] -> [a] -> [a]

stripIfaceInvisVars :: DynFlags -> [IfaceTyConBinder] -> [IfaceTyConBinder]

stripInvisArgs :: DynFlags -> IfaceTcArgs -> IfaceTcArgs

mkIfaceTySubst :: [(IfLclName, IfaceType)] -> IfaceTySubst

substIfaceTyVar :: IfaceTySubst -> IfLclName -> IfaceType

substIfaceTcArgs :: IfaceTySubst -> IfaceTcArgs -> IfaceTcArgs

inDomIfaceTySubst :: IfaceTySubst -> IfaceTvBndr -> Bool