Language.Haskell.Syntax.Basic

Documentation

data Boxity

Constructors

Boxed

Unboxed

isBoxed :: Boxity -> Bool

type SumWidth = Int

The width of an unboxed sum

type ConTag = Int

A *one-index* constructor tag

Type of the tags associated with each constructor possibility or superclass selector

newtype FieldLabelString

Field labels are just represented as strings; they are not necessarily unique (even within a module)

Constructors

FieldLabelString

Fields

field_label :: FastString

data Role

See Note [Roles] in GHC.Core.Coercion

Order of constructors matters: the Ord instance coincides with the *super*typing relation on roles.

Constructors

Nominal

Representational

Phantom

data HsBang

Haskell Bang

Bangs on data constructor arguments written by the user.

(HsBang SrcUnpack SrcLazy) and (HsBang SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsBang NoSrcUnpack SrcLazy)

HsSrcBang is a wrapper around this, associating it with a SourceText as written by the user. In the AST, the SourceText is hidden inside the extension point XBangTy.

Constructors

HsBang SrcUnpackedness SrcStrictness

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie ~

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{-# UNPACK #-} specified

SrcNoUnpack
{-# NOUNPACK #-} specified

NoSrcUnpack
no unpack pragma

data LexicalFixity

Captures the fixity of declarations as they are parsed. This is not necessarily the same as the fixity declaration, as the normal fixity may be overridden using parens or backticks.

Constructors

Prefix

Infix

data FixityDirection

Constructors

InfixL

InfixR

InfixN

data Fixity

Constructors

Fixity Int FixityDirection