Language.Haskell.Syntax.Binds

Documentation

type HsLocalBinds id = HsLocalBindsLR id id

Haskell Local Bindings

type LHsLocalBinds id = XRec id (HsLocalBinds id)

Located Haskell local bindings

data HsLocalBindsLR idL idR

Haskell Local Bindings with separate Left and Right identifier types

Bindings in a 'let' expression or a 'where' clause

Constructors

HsValBinds (XHsValBinds idL idR) (HsValBindsLR idL idR)
Haskell Value Bindings

HsIPBinds (XHsIPBinds idL idR) (HsIPBinds idR)
Haskell Implicit Parameter Bindings

EmptyLocalBinds (XEmptyLocalBinds idL idR)
Empty Local Bindings

XHsLocalBindsLR !(XXHsLocalBindsLR idL idR)

type LHsLocalBindsLR idL idR = XRec idL (HsLocalBindsLR idL idR)

type HsValBinds id = HsValBindsLR id id

Haskell Value Bindings

data HsValBindsLR idL idR

Haskell Value bindings with separate Left and Right identifier types (not implicit parameters) Used for both top level and nested bindings May contain pattern synonym bindings

Constructors

ValBinds (XValBinds idL idR) (LHsBindsLR idL idR) [LSig idR]
Value Bindings In

Before renaming RHS; idR is always RdrName Not dependency analysed Recursive by default

XValBindsLR !(XXValBindsLR idL idR)
Value Bindings Out

After renaming RHS; idR can be Name or Id Dependency analysed, later bindings in the list may depend on earlier ones.

type LHsBind id = LHsBindLR id id

Located Haskell Binding

type LHsBinds id = LHsBindsLR id id

Located Haskell Bindings

type HsBind id = HsBindLR id id

Haskell Binding

type LHsBindsLR idL idR = Bag (LHsBindLR idL idR)

Located Haskell Bindings with separate Left and Right identifier types

type LHsBindLR idL idR = XRec idL (HsBindLR idL idR)

Located Haskell Binding with separate Left and Right identifier types

data HsBindLR idL idR

Haskell Binding with separate Left and Right id's

Constructors

FunBind
Function-like Binding

FunBind is used for both functions f x = e and variables f = x -> e and strict variables !x = x + 1

Reason 1: Special case for type inference: see tcMonoBinds.

Reason 2: Instance decls can only have FunBinds, which is convenient. If you change this, you'll need to change e.g. rnMethodBinds

But note that the form f :: a->a = ... parses as a pattern binding, just like (f :: a -> a) = ...

Strict bindings have their strictness recorded in the SrcStrictness of their MatchContext. See Note [FunBind vs PatBind] for details about the relationship between FunBind and PatBind.

AnnKeywordIds

AnnFunId, attached to each element of fun_matches

AnnEqual,AnnWhere, AnnOpen,AnnClose,

Fields

fun_ext :: XFunBind idL idR

After the renamer (but before the type-checker), this contains the locally-bound free variables of this defn. See Note [Bind free vars]

After the type-checker, this contains a coercion from the type of the MatchGroup to the type of the Id. Example:

     f :: Int -> forall a. a -> a
     f x y = y

Then the MatchGroup will have type (Int -> a' -> a') (with a free type variable a'). The coercion will take a CoreExpr of this type and convert it to a CoreExpr of type Int -> forall a'. a' -> a' Notice that the coercion captures the free a'.

fun_id :: LIdP idL

fun_matches :: MatchGroup idR (LHsExpr idR)

The payload

fun_tick :: [CoreTickish]

Ticks to put on the rhs, if any

PatBind
Pattern Binding

The pattern is never a simple variable; That case is done by FunBind. See Note [FunBind vs PatBind] for details about the relationship between FunBind and PatBind.

Fields

pat_ext :: XPatBind idL idR

See Note [Bind free vars]

pat_lhs :: LPat idL

pat_rhs :: GRHSs idR (LHsExpr idR)

pat_ticks :: ([CoreTickish], [[CoreTickish]])

Ticks to put on the rhs, if any, and ticks to put on the bound variables.

VarBind
Variable Binding

Dictionary binding and suchlike. All VarBinds are introduced by the type checker

Fields

var_ext :: XVarBind idL idR

var_id :: IdP idL

var_rhs :: LHsExpr idR

Located only for consistency

AbsBinds
Abstraction Bindings

Fields

abs_ext :: XAbsBinds idL idR

abs_tvs :: [TyVar]

abs_ev_vars :: [EvVar]

Includes equality constraints

abs_exports :: [ABExport idL]

AbsBinds only gets used when idL = idR after renaming, but these need to be idL's for the collect... code in HsUtil to have the right type

abs_ev_binds :: [TcEvBinds]

Evidence bindings Why a list? See GHC.Tc.TyCl.Instance Note [Typechecking plan for instance declarations]

abs_binds :: LHsBinds idL

Typechecked user bindings

abs_sig :: Bool

PatSynBind
Patterns Synonym Binding

Fields

(XPatSynBind idL idR)

(PatSynBind idL idR)

AnnKeywordId : AnnPattern, AnnLarrow,AnnEqual, AnnWhere AnnOpen '{',AnnClose '}'

XHsBindsLR !(XXHsBindsLR idL idR)

data ABExport p

Abstraction Bindings Export

Constructors

ABE

Fields

abe_ext :: XABE p

abe_poly :: IdP p

Any INLINE pragma is attached to this Id

abe_mono :: IdP p

abe_wrap :: HsWrapper

See Note [ABExport wrapper] Shape: (forall abs_tvs. abs_ev_vars => abe_mono) ~ abe_poly

abe_prags :: TcSpecPrags

SPECIALISE pragmas

XABExport !(XXABExport p)

data PatSynBind idL idR

AnnKeywordId : AnnPattern, AnnEqual,AnnLarrow, AnnWhere,AnnOpen '{', AnnClose '}',

Pattern Synonym binding

Constructors

PSB

Fields

psb_ext :: XPSB idL idR

Post renaming, FVs. See Note [Bind free vars]

psb_id :: LIdP idL

Name of the pattern synonym

psb_args :: HsPatSynDetails idR

Formal parameter names

psb_def :: LPat idR

Right-hand side

psb_dir :: HsPatSynDir idR

Directionality

XPatSynBind !(XXPatSynBind idL idR)

data HsIPBinds id

Haskell Implicit Parameter Bindings

Constructors

IPBinds (XIPBinds id) [LIPBind id]

XHsIPBinds !(XXHsIPBinds id)

type LIPBind id = XRec id (IPBind id)

Located Implicit Parameter Binding

May have AnnKeywordId : AnnSemi when in a list

data IPBind id

Implicit parameter bindings.

These bindings start off as (Left "x") in the parser and stay that way until after type-checking when they are replaced with (Right d), where "d" is the name of the dictionary holding the evidence for the implicit parameter.

AnnKeywordId : AnnEqual

Constructors

IPBind (XCIPBind id) (Either (XRec id HsIPName) (IdP id)) (LHsExpr id)

XIPBind !(XXIPBind id)

type LSig pass = XRec pass (Sig pass)

Located Signature

data Sig pass

Signatures and pragmas

Constructors

TypeSig (XTypeSig pass) [LIdP pass] (LHsSigWcType pass)
An ordinary type signature

f :: Num a => a -> a

After renaming, this list of Names contains the named wildcards brought into scope by this signature. For a signature _ -> _a -> Bool, the renamer will leave the unnamed wildcard _ untouched, and the named wildcard _a is then replaced with fresh meta vars in the type. Their names are stored in the type signature that brought them into scope, in this third field to be more specific.

AnnKeywordId : AnnDcolon, AnnComma

PatSynSig (XPatSynSig pass) [LIdP pass] (LHsSigType pass)
A pattern synonym type signature

pattern Single :: () => (Show a) => a -> [a]

AnnKeywordId : AnnPattern, AnnDcolon,AnnForall AnnDot,AnnDarrow

ClassOpSig (XClassOpSig pass) Bool [LIdP pass] (LHsSigType pass)
A signature for a class method False: ordinary class-method signature True: generic-default class method signature e.g. class C a where op :: a -> a -- Ordinary default op :: Eq a => a -> a -- Generic default No wildcards allowed here

AnnKeywordId : AnnDefault, AnnDcolon

IdSig (XIdSig pass) Id
A type signature in generated code, notably the code generated for record selectors. We simply record the desired Id itself, replete with its name, type and IdDetails. Otherwise it's just like a type signature: there should be an accompanying binding

FixSig (XFixSig pass) (FixitySig pass)
An ordinary fixity declaration

    infixl 8 ***

AnnKeywordId : AnnInfix, AnnVal

InlineSig (XInlineSig pass) (LIdP pass) InlinePragma
An inline pragma

{#- INLINE f #-}

AnnKeywordId : AnnOpen '{-# INLINE' and '[', AnnClose,AnnOpen, AnnVal,AnnTilde, AnnClose

SpecSig (XSpecSig pass) (LIdP pass) [LHsSigType pass] InlinePragma
A specialisation pragma

{-# SPECIALISE f :: Int -> Int #-}

AnnKeywordId : AnnOpen, AnnOpen '{-# SPECIALISE' and '[', AnnTilde, AnnVal, AnnClose ']' and '#-}', AnnDcolon

SpecInstSig (XSpecInstSig pass) SourceText (LHsSigType pass)
A specialisation pragma for instance declarations only

{-# SPECIALISE instance Eq [Int] #-}

(Class tys); should be a specialisation of the current instance declaration

AnnKeywordId : AnnOpen, AnnInstance,AnnClose

MinimalSig (XMinimalSig pass) SourceText (LBooleanFormula (LIdP pass))
A minimal complete definition pragma

{-# MINIMAL a | (b, c | (d | e)) #-}

AnnKeywordId : AnnOpen, AnnVbar,AnnComma, AnnClose

SCCFunSig (XSCCFunSig pass) SourceText (LIdP pass) (Maybe (XRec pass StringLiteral))
A "set cost centre" pragma for declarations

{-# SCC funName #-}

or

{-# SCC funName "cost_centre_name" #-}

CompleteMatchSig (XCompleteMatchSig pass) SourceText (XRec pass [LIdP pass]) (Maybe (LIdP pass))
A complete match pragma

{-# COMPLETE C, D [:: T] #-}

Used to inform the pattern match checker about additional complete matchings which, for example, arise from pattern synonym definitions.

XSig !(XXSig pass)

type LFixitySig pass = XRec pass (FixitySig pass)

Located Fixity Signature

data FixitySig pass

Fixity Signature

Constructors

FixitySig (XFixitySig pass) [LIdP pass] Fixity

XFixitySig !(XXFixitySig pass)

data TcSpecPrags

Type checker Specialisation Pragmas

TcSpecPrags conveys SPECIALISE pragmas from the type checker to the desugarer

Constructors

IsDefaultMethod
Super-specialised: a default method should be macro-expanded at every call site

SpecPrags [LTcSpecPrag]

type LTcSpecPrag = Located TcSpecPrag

Located Type checker Specification Pragmas

data TcSpecPrag

Type checker Specification Pragma

Constructors

SpecPrag Id HsWrapper InlinePragma
The Id to be specialised, a wrapper that specialises the polymorphic function, and inlining spec for the specialised function

noSpecPrags :: TcSpecPrags

hasSpecPrags :: TcSpecPrags -> Bool

isDefaultMethod :: TcSpecPrags -> Bool

isFixityLSig :: forall p. UnXRec p => LSig p -> Bool

isTypeLSig :: forall p. UnXRec p => LSig p -> Bool

isSpecLSig :: forall p. UnXRec p => LSig p -> Bool

isSpecInstLSig :: forall p. UnXRec p => LSig p -> Bool

isPragLSig :: forall p. UnXRec p => LSig p -> Bool

isInlineLSig :: forall p. UnXRec p => LSig p -> Bool

isMinimalLSig :: forall p. UnXRec p => LSig p -> Bool

isSCCFunSig :: forall p. UnXRec p => LSig p -> Bool

isCompleteMatchSig :: forall p. UnXRec p => LSig p -> Bool

hsSigDoc :: Sig name -> SDoc

type HsPatSynDetails pass = HsConDetails Void (LIdP pass) [RecordPatSynField pass]

Haskell Pattern Synonym Details

data RecordPatSynField pass

Record Pattern Synonym Field

Constructors

RecordPatSynField

Fields

recordPatSynField :: FieldOcc pass

Field label visible in rest of the file

recordPatSynPatVar :: LIdP pass

Filled in by renamer, the name used internally by the pattern

data HsPatSynDir id

Haskell Pattern Synonym Direction

Constructors

Unidirectional

ImplicitBidirectional

ExplicitBidirectional (MatchGroup id (LHsExpr id))