Language.Haskell.Syntax.Concrete

Description

Bits of concrete syntax (tokens, layout).

Documentation

type LHsToken tok p = XRec p (HsToken tok)

type LHsUniToken tok utok p = XRec p (HsUniToken tok utok)

data HsToken (tok :: Symbol)

A token stored in the syntax tree. For example, when parsing a let-expression, we store HsToken "let" and HsToken "in". The locations of those tokens can be used to faithfully reproduce (exactprint) the original program text.

Constructors

HsTok

data HsUniToken (tok :: Symbol) (utok :: Symbol)

With UnicodeSyntax, there might be multiple ways to write the same token. For example an arrow could be either -> or →. This choice must be recorded in order to exactprint such tokens, so instead of HsToken "->" we introduce HsUniToken "->" "→".

See also IsUnicodeSyntax in GHC.Parser.Annotation; we do not use here to avoid a dependency.

Constructors

HsNormalTok

HsUnicodeTok

data LayoutInfo pass

Layout information for declarations.

Constructors

ExplicitBraces !(LHsToken "{" pass) !(LHsToken "}" pass)
Explicit braces written by the user.

class C a where { foo :: a; bar :: a }

VirtualBraces
Virtual braces inserted by the layout algorithm.

class C a where
  foo :: a
  bar :: a

Fields

!Int

Layout column (indentation level, begins at 1)

NoLayoutInfo
Empty or compiler-generated blocks do not have layout information associated with them.