Language.Haskell.Syntax.Expr

Description

Abstract Haskell syntax for expressions.

Documentation

type LFieldLabelStrings p = XRec p (FieldLabelStrings p)

RecordDotSyntax field updates

newtype FieldLabelStrings p

Constructors

FieldLabelStrings [XRec p (DotFieldOcc p)]

type RecProj p arg = HsFieldBind (LFieldLabelStrings p) arg

type LHsRecProj p arg = XRec p (RecProj p arg)

type RecUpdProj p = RecProj p (LHsExpr p)

type LHsRecUpdProj p = XRec p (RecUpdProj p)

data LHsRecUpdFields p where

Haskell Record Update Fields.

Constructors

RegularRecUpdFields
A regular (non-overloaded) record update.

Fields

:: forall p. { xRecUpdFields :: XLHsRecUpdLabels p

, recUpdFields :: [LHsRecUpdField p p]

} -> LHsRecUpdFields p

OverloadedRecUpdFields
An overloaded record update.

Fields

:: forall p. { xOLRecUpdFields :: XLHsOLRecUpdLabels p

, olRecUpdFields :: [LHsRecUpdProj p]

} -> LHsRecUpdFields p

Expressions proper

type LHsExpr p = XRec p (HsExpr p)

Located Haskell Expression

type family SyntaxExpr p

Syntax Expression

SyntaxExpr is represents the function used in interpreting rebindable syntax. In the parser, we have no information to supply; in the renamer, we have the name of the function (but see Note [Monad fail : Rebindable syntax, overloaded strings] for a wrinkle) and in the type-checker we have a more elaborate structure SyntaxExprTc.

In some contexts, rebindable syntax is not implemented, and so we have constructors to represent that possibility in both the renamer and typechecker instantiations.

E.g. (>>=) is filled in before the renamer by the appropriate Name for (>>=), and then instantiated by the type checker with its type args etc

data HsExpr p

A Haskell expression.

Constructors

HsVar (XVar p) (LIdP p)
Variable See Note [Located RdrNames]

HsUnboundVar (XUnboundVar p) RdrName
Unbound variable; also used for "holes" (_ or _x). Turned from HsVar to HsUnboundVar by the renamer, when it finds an out-of-scope variable or hole. The (XUnboundVar p) field becomes an HoleExprRef after typechecking; this is where the erroring expression will be written after solving. See Note [Holes] in GHC.Tc.Types.Constraint.

HsOverLabel (XOverLabel p) FastString
Overloaded label (Note [Overloaded labels] in GHC.OverloadedLabels)

HsIPVar (XIPVar p) HsIPName
Implicit parameter (not in use after typechecking)

HsOverLit (XOverLitE p) (HsOverLit p)
Overloaded literals

HsLit (XLitE p) (HsLit p)
Simple (non-overloaded) literals

HsLam
Lambda, Lambda-case, and Lambda-cases

Fields

(XLam p)

HsLamVariant

Tells whether this is for lambda, case, or cases

(MatchGroup p (LHsExpr p))

LamSingle: one match of arity >= 1 LamCase: many arity-1 matches LamCases: many matches of uniform arity >= 1

HsApp (XApp p) (LHsExpr p) (LHsExpr p)
Application

HsAppType (XAppTypeE p) (LHsExpr p) (LHsWcType (NoGhcTc p))
Visible type application

Explicit type argument; e.g f @Int x y NB: Has wildcards, but no implicit quantification

OpApp (XOpApp p) (LHsExpr p) (LHsExpr p) (LHsExpr p)
Operator applications: NB Bracketed ops such as (+) come out as Vars.

NegApp (XNegApp p) (LHsExpr p) (SyntaxExpr p)
Negation operator. Contains the negated expression and the name of negate

HsPar (XPar p) (LHsExpr p)
Parenthesised expr; see Note [Parens in HsSyn]

SectionL (XSectionL p) (LHsExpr p) (LHsExpr p)

SectionR (XSectionR p) (LHsExpr p) (LHsExpr p)

ExplicitTuple (XExplicitTuple p) [HsTupArg p] Boxity
Used for explicit tuples and sections thereof

ExplicitSum (XExplicitSum p) ConTag SumWidth (LHsExpr p)
Used for unboxed sum types

HsCase (XCase p) (LHsExpr p) (MatchGroup p (LHsExpr p))

HsIf (XIf p) (LHsExpr p) (LHsExpr p) (LHsExpr p)

HsMultiIf (XMultiIf p) [LGRHS p (LHsExpr p)]
Multi-way if

HsLet (XLet p) (HsLocalBinds p) (LHsExpr p)
let(rec)

HsDo (XDo p) HsDoFlavour (XRec p [ExprLStmt p])

ExplicitList (XExplicitList p) [LHsExpr p]
Syntactic list: [a,b,c,...]

RecordCon
Record construction

Fields

rcon_ext :: XRecordCon p

rcon_con :: XRec p (ConLikeP p)

rcon_flds :: HsRecordBinds p

RecordUpd
Record update

Fields

rupd_ext :: XRecordUpd p

rupd_expr :: LHsExpr p

rupd_flds :: LHsRecUpdFields p

HsGetField
Record field selection e.g z.x.

Fields

gf_ext :: XGetField p

gf_expr :: LHsExpr p

gf_field :: XRec p (DotFieldOcc p)

HsProjection
Record field selector. e.g. (.x) or (.x.y)

This case only arises when the OverloadedRecordDot langauge extensions is enabled. See Note [Record selectors in the AST].

Fields

proj_ext :: XProjection p

proj_flds :: NonEmpty (DotFieldOcc p)

ExprWithTySig (XExprWithTySig p) (LHsExpr p) (LHsSigWcType (NoGhcTc p))
Expression with an explicit type signature. e :: type

ArithSeq (XArithSeq p) (Maybe (SyntaxExpr p)) (ArithSeqInfo p)
Arithmetic sequence

HsTypedBracket (XTypedBracket p) (LHsExpr p)

HsUntypedBracket (XUntypedBracket p) (HsQuote p)

HsTypedSplice (XTypedSplice p) (LHsExpr p)

HsUntypedSplice (XUntypedSplice p) (HsUntypedSplice p)

HsProc (XProc p) (LPat p) (LHsCmdTop p)
proc notation for Arrows

HsStatic (XStatic p) (LHsExpr p)

HsPragE (XPragE p) (HsPragE p) (LHsExpr p)

HsEmbTy (XEmbTy p) (LHsWcType (NoGhcTc p))

HsForAll (XForAll p) (HsForAllTelescope p) (LHsExpr p)
Forall-types forall tvs. t and forall tvs -> t. Used with RequiredTypeArguments, e.g. fn (forall a. Proxy a). See Note [Types in terms]

HsQual (XQual p) (XRec p [LHsExpr p]) (LHsExpr p)

HsFunArr (XFunArr p) (HsArrowOf (LHsExpr p) p) (LHsExpr p) (LHsExpr p)
Function types a -> b. Used with RequiredTypeArguments, e.g. fn (Int -> Bool). See Note [Types in terms]

XExpr !(XXExpr p)

data DotFieldOcc p

Constructors

DotFieldOcc

Fields

dfoExt :: XCDotFieldOcc p

dfoLabel :: XRec p FieldLabelString

XDotFieldOcc !(XXDotFieldOcc p)

data HsPragE p

A pragma, written as {-# ... #-}, that may appear within an expression.

Constructors

HsPragSCC (XSCC p) StringLiteral

XHsPragE !(XXPragE p)

type LHsTupArg id = XRec id (HsTupArg id)

Located Haskell Tuple Argument

HsTupArg is used for tuple sections (,a,) is represented by ExplicitTuple [Missing ty1, Present a, Missing ty3] Which in turn stands for (x:ty1 y:ty2. (x,a,y))

data HsTupArg id

Haskell Tuple Argument

Constructors

Present (XPresent id) (LHsExpr id)
The argument

Missing (XMissing id)
The argument is missing, but this is its type

XTupArg !(XXTupArg id)
Extension point; see Note [Trees That Grow] in Language.Haskell.Syntax.Extension

data HsLamVariant

Which kind of lambda case are we dealing with?

Constructors

LamSingle
`p -> e`

LamCase
`case pi -> ei `

LamCases
`cases psi -> ei`

type LHsCmd id = XRec id (HsCmd id)

Located Haskell Command (for arrow syntax)

data HsCmd id

Haskell Command (e.g. a "statement" in an Arrow proc block)

Constructors

HsCmdArrApp (XCmdArrApp id) (LHsExpr id) (LHsExpr id) HsArrAppType Bool

HsCmdArrForm (XCmdArrForm id) (LHsExpr id) LexicalFixity [LHsCmdTop id]

HsCmdApp (XCmdApp id) (LHsCmd id) (LHsExpr id)

HsCmdLam (XCmdLamCase id) HsLamVariant (MatchGroup id (LHsCmd id))
Lambda-case

HsCmdPar (XCmdPar id) (LHsCmd id)

HsCmdCase (XCmdCase id) (LHsExpr id) (MatchGroup id (LHsCmd id))

HsCmdIf (XCmdIf id) (SyntaxExpr id) (LHsExpr id) (LHsCmd id) (LHsCmd id)

HsCmdLet (XCmdLet id) (HsLocalBinds id) (LHsCmd id)

HsCmdDo (XCmdDo id) (XRec id [CmdLStmt id])

XCmd !(XXCmd id)

data HsArrAppType

Haskell arrow application type.

Constructors

HsHigherOrderApp
First order arrow application -<

HsFirstOrderApp
Higher order arrow application -<<

type LHsCmdTop p = XRec p (HsCmdTop p)

Top-level command, introducing a new arrow. This may occur inside a proc (where the stack is empty) or as an argument of a command-forming operator.

Located Haskell Top-level Command

data HsCmdTop p

Haskell Top-level Command

Constructors

HsCmdTop (XCmdTop p) (LHsCmd p)

XCmdTop !(XXCmdTop p)

type HsRecordBinds p = HsRecFields p (LHsExpr p)

Haskell Record Bindings

data MatchGroup p body

Constructors

MG

Fields

mg_ext :: XMG p body

mg_alts :: XRec p [LMatch p body]

XMatchGroup !(XXMatchGroup p body)

type LMatch id body = XRec id (Match id body)

Located Match

data Match p body

Constructors

Match

Fields

m_ext :: XCMatch p body

m_ctxt :: HsMatchContext (LIdP (NoGhcTc p))

m_pats :: XRec p [LPat p]

m_grhss :: GRHSs p body

XMatch !(XXMatch p body)

isInfixMatch :: Match id body -> Bool

data GRHSs p body

Guarded Right-Hand Sides

GRHSs are used both for pattern bindings and for Matches

Constructors

GRHSs

Fields

grhssExt :: XCGRHSs p body

grhssGRHSs :: [LGRHS p body]

Guarded RHSs

grhssLocalBinds :: HsLocalBinds p

The where clause

XGRHSs !(XXGRHSs p body)

type LGRHS id body = XRec id (GRHS id body)

Located Guarded Right-Hand Side

data GRHS p body

Guarded Right Hand Side.

Constructors

GRHS (XCGRHS p body) [GuardLStmt p] body

XGRHS !(XXGRHS p body)

type LStmt id body = XRec id (StmtLR id id body)

Located do block Statement

type LStmtLR idL idR body = XRec idL (StmtLR idL idR body)

Located Statement with separate Left and Right id's

type Stmt id body = StmtLR id id body

do block Statement

type CmdLStmt id = LStmt id (LHsCmd id)

Command Located Statement

type CmdStmt id = Stmt id (LHsCmd id)

Command Statement

type ExprLStmt id = LStmt id (LHsExpr id)

Expression Located Statement

type ExprStmt id = Stmt id (LHsExpr id)

Expression Statement

type GuardLStmt id = LStmt id (LHsExpr id)

Guard Located Statement

type GuardStmt id = Stmt id (LHsExpr id)

Guard Statement

type GhciLStmt id = LStmt id (LHsExpr id)

Ghci Located Statement

type GhciStmt id = Stmt id (LHsExpr id)

Ghci Statement

data StmtLR idL idR body

Constructors

LastStmt (XLastStmt idL idR body) body (Maybe Bool) (SyntaxExpr idR)

BindStmt

Fields

(XBindStmt idL idR body)

Post renaming has optional fail and bind / (>>=) operator. Post typechecking, also has multiplicity of the argument and the result type of the function passed to bind; that is, (P, S) in (>>=) :: Q -> (R % P -> S) -> T See Note [The type of bind in Stmts]

(LPat idL)

body

BodyStmt (XBodyStmt idL idR body) body (SyntaxExpr idR) (SyntaxExpr idR)

LetStmt (XLetStmt idL idR body) (HsLocalBindsLR idL idR)

ParStmt (XParStmt idL idR body) [ParStmtBlock idL idR] (HsExpr idR) (SyntaxExpr idR)

TransStmt

Fields

trS_ext :: XTransStmt idL idR body

trS_form :: TransForm

trS_stmts :: [ExprLStmt idL]

trS_bndrs :: [(IdP idR, IdP idR)]

trS_using :: LHsExpr idR

trS_by :: Maybe (LHsExpr idR)

trS_ret :: SyntaxExpr idR

trS_bind :: SyntaxExpr idR

trS_fmap :: HsExpr idR

RecStmt

Fields

recS_ext :: XRecStmt idL idR body

recS_stmts :: XRec idR [LStmtLR idL idR body]

recS_later_ids :: [IdP idR]

recS_rec_ids :: [IdP idR]

recS_bind_fn :: SyntaxExpr idR

recS_ret_fn :: SyntaxExpr idR

recS_mfix_fn :: SyntaxExpr idR

XStmtLR !(XXStmtLR idL idR body)

data TransForm

Constructors

ThenForm

GroupForm

data ParStmtBlock idL idR

Parenthesised Statement Block

Constructors

ParStmtBlock (XParStmtBlock idL idR) [ExprLStmt idL] [IdP idR] (SyntaxExpr idR)

XParStmtBlock !(XXParStmtBlock idL idR)

type FailOperator id = Maybe (SyntaxExpr id)

The fail operator

This is used for `.. <-` "bind statements" in do notation, including non-monadic "binds" in applicative.

The fail operator is 'Just expr' if it potentially fail monadically. if the pattern match cannot fail, or shouldn't fail monadically (regular incomplete pattern exception), it is Nothing.

See Note [Monad fail : Rebindable syntax, overloaded strings] for the type of expression in the Just case, and why it is so.

See Note [Failing pattern matches in Stmts] for which contexts for 'BindStmt's should use the monadic fail and which shouldn't.

data HsUntypedSplice id

Haskell Splice

Constructors

HsUntypedSpliceExpr (XUntypedSpliceExpr id) (LHsExpr id)

HsQuasiQuote (XQuasiQuote id) (IdP id) (XRec id FastString)

XUntypedSplice !(XXUntypedSplice id)

data HsQuote p

Haskell (Untyped) Quote = Expr + Pat + Type + Var

Constructors

ExpBr (XExpBr p) (LHsExpr p)

PatBr (XPatBr p) (LPat p)

DecBrL (XDecBrL p) [LHsDecl p]

DecBrG (XDecBrG p) (HsGroup p)

TypBr (XTypBr p) (LHsType p)

VarBr (XVarBr p) Bool (LIdP p)

XQuote !(XXQuote p)

data ArithSeqInfo id

Arithmetic Sequence Information

Constructors

From (LHsExpr id)

FromThen (LHsExpr id) (LHsExpr id)

FromTo (LHsExpr id) (LHsExpr id)

FromThenTo (LHsExpr id) (LHsExpr id) (LHsExpr id)

data HsMatchContext fn

Haskell Match Context

Context of a pattern match. This is more subtle than it would seem. See Note [FunBind vs PatBind].

Constructors

FunRhs
A pattern matching on an argument of a function binding

Fields

mc_fun :: fn

function binder of f See Note [mc_fun field of FunRhs] See #20415 for a long discussion about this field

mc_fixity :: LexicalFixity

fixing of f

mc_strictness :: SrcStrictness

was f banged? See Note [FunBind vs PatBind]

mc_an :: XFunRhs

CaseAlt
Patterns and guards in a case alternative

LamAlt HsLamVariant
Patterns and guards in @, case and cases@

IfAlt
Guards of a multi-way if alternative

ArrowMatchCtxt HsArrowMatchContext
A pattern match inside arrow notation

PatBindRhs
A pattern binding eg [y] <- e = e

PatBindGuards
Guards of pattern bindings, e.g., (Just b) | Just _ <- x = e | otherwise = e'

RecUpd
Record update [used only in GHC.HsToCore.Expr to tell matchWrapper what sort of runtime error message to generate]

StmtCtxt (HsStmtContext fn)
Pattern of a do-stmt, list comprehension, pattern guard, etc

ThPatSplice
A Template Haskell pattern splice

ThPatQuote
A Template Haskell pattern quotation [p| (a,b) |]

PatSyn
A pattern synonym declaration

LazyPatCtx
An irrefutable pattern

data HsStmtContext fn

Haskell Statement Context.

Constructors

HsDoStmt HsDoFlavour
Context for HsDo (do-notation and comprehensions)

PatGuard (HsMatchContext fn)
Pattern guard for specified thing

ParStmtCtxt (HsStmtContext fn)
A branch of a parallel stmt

TransStmtCtxt (HsStmtContext fn)
A branch of a transform stmt

ArrowExpr
do-notation in an arrow-command context

data HsArrowMatchContext

Haskell arrow match context.

Constructors

ProcExpr
A proc expression

ArrowCaseAlt
A case alternative inside arrow notation

ArrowLamAlt HsLamVariant
A , case or cases alternative inside arrow notation

data HsDoFlavour

Constructors

DoExpr (Maybe ModuleName)

ModuleName.

do { ... }

MDoExpr (Maybe ModuleName)

ModuleName.

mdo { ... } ie recursive do-expression

GhciStmtCtxt
A command-line Stmt in GHCi pat <- rhs

ListComp

MonadComp

qualifiedDoModuleName_maybe :: HsStmtContext fn -> Maybe ModuleName

isPatSynCtxt :: HsMatchContext fn -> Bool

isComprehensionContext :: HsStmtContext fn -> Bool

isDoComprehensionContext :: HsDoFlavour -> Bool

isMonadStmtContext :: HsStmtContext fn -> Bool

Is this a monadic context?

isMonadDoStmtContext :: HsDoFlavour -> Bool

isMonadCompContext :: HsStmtContext fn -> Bool

isMonadDoCompContext :: HsDoFlavour -> Bool