Language.Haskell.Syntax.Extension

Documentation

data NoExtField

A placeholder type for TTG extension points that are not currently unused to represent any particular value.

This should not be confused with NoExtCon, which are found in unused extension constructors and therefore should never be inhabited. In contrast, NoExtField is used in extension points (e.g., as the field of some constructor), so it must have an inhabitant to construct AST passes that manipulate fields with that extension point as their type.

Constructors

NoExtField

noExtField :: NoExtField

Used when constructing a term with an unused extension point.

data NoExtCon

Used in TTG extension constructors that have yet to be extended with anything. If an extension constructor has NoExtCon as its field, it is not intended to ever be constructed anywhere, and any function that consumes the extension constructor can eliminate it by way of noExtCon.

This should not be confused with NoExtField, which are found in unused extension points (not constructors) and therefore can be inhabited.

noExtCon :: NoExtCon -> a

Eliminate a NoExtCon. Much like absurd.

type family XRec p a = r | r -> a

GHC's L prefixed variants wrap their vanilla variant in this type family, to add SrcLoc info via Located. Other passes than GhcPass not interested in location information can define this as type instance XRec NoLocated a = a. See Note [XRec and SrcSpans in the AST]

type family Anno a = b

class UnXRec p where

We can strip off the XRec to access the underlying data. See Note [XRec and SrcSpans in the AST]

Methods

unXRec :: XRec p a -> a

class MapXRec p where

We can map over the underlying type contained in an XRec while preserving the annotation as is.

Methods

mapXRec :: Anno a ~ Anno b => (a -> b) -> XRec p a -> XRec p b

class WrapXRec p a where

The trivial wrapper that carries no additional information See Note [XRec and SrcSpans in the AST]

Methods

wrapXRec :: a -> XRec p a

type family IdP p

Maps the "normal" id type for a given pass

type LIdP p = XRec p (IdP p)

type family XHsValBinds x x'

type family XHsIPBinds x x'

type family XEmptyLocalBinds x x'

type family XXHsLocalBindsLR x x'

type family XValBinds x x'

type family XXValBindsLR x x'

type family XFunBind x x'

type family XPatBind x x'

type family XVarBind x x'

type family XAbsBinds x x'

type family XPatSynBind x x'

type family XXHsBindsLR x x'

type family XABE x

type family XXABExport x

type family XPSB x x'

type family XXPatSynBind x x'

type family XIPBinds x

type family XXHsIPBinds x

type family XCIPBind x

type family XXIPBind x

type family XTypeSig x

type family XPatSynSig x

type family XClassOpSig x

type family XIdSig x

type family XFixSig x

type family XInlineSig x

type family XSpecSig x

type family XSpecInstSig x

type family XMinimalSig x

type family XSCCFunSig x

type family XCompleteMatchSig x

type family XXSig x

type family XFixitySig x

type family XXFixitySig x

type family XStandaloneKindSig x

type family XXStandaloneKindSig x

type family XTyClD x

type family XInstD x

type family XDerivD x

type family XValD x

type family XSigD x

type family XKindSigD x

type family XDefD x

type family XForD x

type family XWarningD x

type family XAnnD x

type family XRuleD x

type family XSpliceD x

type family XDocD x

type family XRoleAnnotD x

type family XXHsDecl x

type family XCHsGroup x

type family XXHsGroup x

type family XSpliceDecl x

type family XXSpliceDecl x

type family XFamDecl x

type family XSynDecl x

type family XDataDecl x

type family XClassDecl x

type family XXTyClDecl x

type family XCFunDep x

type family XXFunDep x

type family XCTyClGroup x

type family XXTyClGroup x

type family XNoSig x

type family XCKindSig x

type family XTyVarSig x

type family XXFamilyResultSig x

type family XCFamilyDecl x

type family XXFamilyDecl x

type family XCHsDataDefn x

type family XXHsDataDefn x

type family XCHsDerivingClause x

type family XXHsDerivingClause x

type family XDctSingle x

type family XDctMulti x

type family XXDerivClauseTys x

type family XConDeclGADT x

type family XConDeclH98 x

type family XXConDecl x

type family XCFamEqn x r

type family XXFamEqn x r

type family XCTyFamInstDecl x

type family XXTyFamInstDecl x

type family XCClsInstDecl x

type family XXClsInstDecl x

type family XClsInstD x

type family XDataFamInstD x

type family XTyFamInstD x

type family XXInstDecl x

type family XCDerivDecl x

type family XXDerivDecl x

type family XStockStrategy x

type family XAnyClassStrategy x

type family XNewtypeStrategy x

type family XViaStrategy x

type family XCDefaultDecl x

type family XXDefaultDecl x

type family XForeignImport x

type family XForeignExport x

type family XXForeignDecl x

type family XCRuleDecls x

type family XXRuleDecls x

type family XHsRule x

type family XXRuleDecl x

type family XCRuleBndr x

type family XRuleBndrSig x

type family XXRuleBndr x

type family XWarnings x

type family XXWarnDecls x

type family XWarning x

type family XXWarnDecl x

type family XHsAnnotation x

type family XXAnnDecl x

type family XCRoleAnnotDecl x

type family XXRoleAnnotDecl x

type family XCInjectivityAnn x

type family XXInjectivityAnn x

type family XVar x

type family XUnboundVar x

type family XConLikeOut x

type family XRecFld x

type family XOverLabel x

type family XIPVar x

type family XOverLitE x

type family XLitE x

type family XLam x

type family XLamCase x

type family XApp x

type family XAppTypeE x

type family XOpApp x

type family XNegApp x

type family XPar x

type family XSectionL x

type family XSectionR x

type family XExplicitTuple x

type family XExplicitSum x

type family XCase x

type family XIf x

type family XMultiIf x

type family XLet x

type family XDo x

type family XExplicitList x

type family XRecordCon x

type family XRecordUpd x

type family XGetField x

type family XProjection x

type family XExprWithTySig x

type family XArithSeq x

type family XBracket x

type family XRnBracketOut x

type family XTcBracketOut x

type family XSpliceE x

type family XProc x

type family XStatic x

type family XTick x

type family XBinTick x

type family XPragE x

type family XXExpr x

type family XCHsFieldLabel x

type family XXHsFieldLabel x

type family XSCC x

type family XXPragE x

type family XUnambiguous x

type family XAmbiguous x

type family XXAmbiguousFieldOcc x

type family XPresent x

type family XMissing x

type family XXTupArg x

type family XTypedSplice x

type family XUntypedSplice x

type family XQuasiQuote x

type family XSpliced x

type family XXSplice x

type family XExpBr x

type family XPatBr x

type family XDecBrL x

type family XDecBrG x

type family XTypBr x

type family XVarBr x

type family XTExpBr x

type family XXBracket x

type family XCmdTop x

type family XXCmdTop x

type family XMG x b

type family XXMatchGroup x b

type family XCMatch x b

type family XXMatch x b

type family XCGRHSs x b

type family XXGRHSs x b

type family XCGRHS x b

type family XXGRHS x b

type family XLastStmt x x' b

type family XBindStmt x x' b

type family XApplicativeStmt x x' b

type family XBodyStmt x x' b

type family XLetStmt x x' b

type family XParStmt x x' b

type family XTransStmt x x' b

type family XRecStmt x x' b

type family XXStmtLR x x' b

type family XCmdArrApp x

type family XCmdArrForm x

type family XCmdApp x

type family XCmdLam x

type family XCmdPar x

type family XCmdCase x

type family XCmdLamCase x

type family XCmdIf x

type family XCmdLet x

type family XCmdDo x

type family XCmdWrap x

type family XXCmd x

type family XParStmtBlock x x'

type family XXParStmtBlock x x'

type family XApplicativeArgOne x

type family XApplicativeArgMany x

type family XXApplicativeArg x

type family XHsChar x

type family XHsCharPrim x

type family XHsString x

type family XHsStringPrim x

type family XHsInt x

type family XHsIntPrim x

type family XHsWordPrim x

type family XHsInt64Prim x

type family XHsWord64Prim x

type family XHsInteger x

type family XHsRat x

type family XHsFloatPrim x

type family XHsDoublePrim x

type family XXLit x

type family XOverLit x

type family XXOverLit x

type family XWildPat x

type family XVarPat x

type family XLazyPat x

type family XAsPat x

type family XParPat x

type family XBangPat x

type family XListPat x

type family XTuplePat x

type family XSumPat x

type family XConPat x

type family XViewPat x

type family XSplicePat x

type family XLitPat x

type family XNPat x

type family XNPlusKPat x

type family XSigPat x

type family XCoPat x

type family XXPat x

type family XHsRecField x

type family XHsQTvs x

type family XXLHsQTyVars x

type family XHsOuterImplicit x

type family XHsOuterExplicit x flag

type family XXHsOuterTyVarBndrs x

type family XHsSig x

type family XXHsSigType x

type family XHsWC x b

type family XXHsWildCardBndrs x b

type family XHsPS x

type family XXHsPatSigType x

type family XForAllTy x

type family XQualTy x

type family XTyVar x

type family XAppTy x

type family XAppKindTy x

type family XFunTy x

type family XListTy x

type family XTupleTy x

type family XSumTy x

type family XOpTy x

type family XParTy x

type family XIParamTy x

type family XStarTy x

type family XKindSig x

type family XSpliceTy x

type family XDocTy x

type family XBangTy x

type family XRecTy x

type family XExplicitListTy x

type family XExplicitTupleTy x

type family XTyLit x

type family XWildCardTy x

type family XXType x

type family XHsForAllVis x

type family XHsForAllInvis x

type family XXHsForAllTelescope x

type family XUserTyVar x

type family XKindedTyVar x

type family XXTyVarBndr x

type family XConDeclField x

type family XXConDeclField x

type family XCFieldOcc x

type family XXFieldOcc x

type family XCImportDecl x

type family XXImportDecl x

type family XIEVar x

type family XIEThingAbs x

type family XIEThingAll x

type family XIEThingWith x

type family XIEModuleContents x

type family XIEGroup x

type family XIEDoc x

type family XIEDocNamed x

type family XXIE x

type family NoGhcTc (p :: Type)

See Note [NoGhcTc] in GHC.Hs.Extension. It has to be in this module because it is used like an extension point (in the data definitions of types that should be parameter-agnostic.