Language.Haskell.Syntax.ImpExp

Documentation

type LImportDecl pass = XRec pass (ImportDecl pass)

Located Import Declaration

data ImportDeclQualifiedStyle

If/how an import is qualified.

Constructors

QualifiedPre
qualified appears in prepositive position.

QualifiedPost
qualified appears in postpositive position.

NotQualified
Not qualified.

data IsBootInterface

Indicates whether a module name is referring to a boot interface (hs-boot file) or regular module (hs file). We need to treat boot modules specially when building compilation graphs, since they break cycles. Regular source files and signature files are treated equivalently.

Constructors

NotBoot

IsBoot

data ImportDecl pass

Import Declaration

A single Haskell import declaration.

Constructors

ImportDecl

Fields

ideclExt :: XCImportDecl pass

Locations of keywords like import, qualified, etc. are captured here.

ideclName :: XRec pass ModuleName

Module name.

ideclPkgQual :: ImportDeclPkgQual pass

Package qualifier.

ideclSource :: IsBootInterface

IsBoot <=> {-# SOURCE #-} import

ideclSafe :: Bool

True => safe import

ideclQualified :: ImportDeclQualifiedStyle

If/how the import is qualified.

ideclAs :: Maybe (XRec pass ModuleName)

as Module

ideclImportList :: Maybe (ImportListInterpretation, XRec pass [LIE pass])

Explicit import list (EverythingBut => hiding, names)

XImportDecl !(XXImportDecl pass)

data ImportListInterpretation

Whether the import list is exactly what to import, or whether hiding was used, and therefore everything but what was listed should be imported

Constructors

Exactly

EverythingBut

type LIE pass

Arguments

= XRec pass (IE pass)
When in a list this may have

Located Import or Export

type ExportDoc pass = LHsDoc pass

A docstring attached to an export list item.

data IE pass

Imported or exported entity.

Constructors

IEVar (XIEVar pass) (LIEWrappedName pass) (Maybe (ExportDoc pass))
Imported or exported variable

module Mod ( test )
import Mod ( test )

IEThingAbs (XIEThingAbs pass) (LIEWrappedName pass) (Maybe (ExportDoc pass))
Imported or exported Thing with absent subordinate list

The thing is a Class/Type (can't tell)

module Mod ( Test )
import Mod ( Test )

IEThingAll (XIEThingAll pass) (LIEWrappedName pass) (Maybe (ExportDoc pass))
Imported or exported thing with wildcard subordinate list (e.g. (..))

The thing is a Class/Type and the All refers to methods/constructors

module Mod ( Test(..) )
import Mod ( Test(..) )

IEThingWith (XIEThingWith pass) (LIEWrappedName pass) IEWildcard [LIEWrappedName pass] (Maybe (ExportDoc pass))
Imported or exported thing with explicit subordinate list.

The thing is a Class/Type (can't tell) and the imported or exported things are its children.

module Mod ( Test(f, g) )
import Mod ( Test(f, g) )

IEModuleContents (XIEModuleContents pass) (XRec pass ModuleName)
Export of entire module. Can only occur in export list.

module Mod ( module Mod2 )

IEGroup (XIEGroup pass) Int (LHsDoc pass)
A Haddock section in an export list.

module Mod
  ( -- * Section heading
    ...
  )

IEDoc (XIEDoc pass) (LHsDoc pass)
A bit of unnamed documentation.

module Mod
  ( -- | Documentation
    ...
  )

IEDocNamed (XIEDocNamed pass) String
A reference to a named documentation chunk.

module Mod
  ( -- $chunkName
    ...
  )

XIE !(XXIE pass)

data IEWildcard

Wildcard in an import or export sublist, like the .. in import Mod ( T(Mk1, Mk2, ..) ).

Constructors

NoIEWildcard
no wildcard in this list

IEWildcard Int
wildcard after the given # of items in this list The Int is in the range [0..n], where n is the length of the list.

data IEWrappedName p

A name in an import or export specification which may have adornments. Used primarily for accurate pretty printing of ParsedSource, and API Annotation placement.

Constructors

IEName (XIEName p) (LIdP p)
unadorned name, e.g myFun

IEDefault (XIEDefault p) (LIdP p)
default X (), see Note [Named default declarations] in GHC.Tc.Gen.Default

IEPattern (XIEPattern p) (LIdP p)
pattern X

exactprint: the location of pattern keyword is captured via EpaLocation

IEType (XIEType p) (LIdP p)
type (:+:)

exactprint: the location of type keyword is captured via EpaLocation

XIEWrappedName !(XXIEWrappedName p)

type LIEWrappedName p = XRec p (IEWrappedName p)

Located name with possible adornment