Language.Haskell.Syntax.Module.Name

Documentation

newtype ModuleName

A ModuleName is essentially a simple string, e.g. Data.List.

Constructors

ModuleName FastString

stableModuleNameCmp :: ModuleName -> ModuleName -> Ordering

Compares module names lexically, rather than by their Uniques

moduleNameFS :: ModuleName -> FastString

moduleNameString :: ModuleName -> String

mkModuleName :: String -> ModuleName

mkModuleNameFS :: FastString -> ModuleName

moduleNameSlashes :: ModuleName -> String

Returns the string version of the module name, with dots replaced by slashes.

moduleNameColons :: ModuleName -> String

Returns the string version of the module name, with dots replaced by colons.

parseModuleName :: ReadP ModuleName