Language.Haskell.Syntax.Specificity

ForAllTyFlags

data ForAllTyFlag

ForAllTyFlag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, ForAllTyBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible !Specificity

Required

Bundled Patterns

pattern Specified :: ForAllTyFlag

pattern Inferred :: ForAllTyFlag

data Specificity

Whether an Invisible argument may appear in source Haskell.

Constructors

InferredSpec
the argument may not appear in source Haskell, it is only inferred.

SpecifiedSpec
the argument may appear in source Haskell, but isn't required.

isVisibleForAllTyFlag :: ForAllTyFlag -> Bool

Does this ForAllTyFlag classify an argument that is written in Haskell?

isInvisibleForAllTyFlag :: ForAllTyFlag -> Bool

Does this ForAllTyFlag classify an argument that is not written in Haskell?

isInferredForAllTyFlag :: ForAllTyFlag -> Bool

isSpecifiedForAllTyFlag :: ForAllTyFlag -> Bool

coreTyLamForAllTyFlag :: ForAllTyFlag

The ForAllTyFlag on a (Lam a e) term, where a is a type variable. If you want other ForAllTyFlag, use a cast. See Note [Required foralls in Core] in GHC.Core.TyCo.Rep