Language.Haskell.Syntax.Type

Documentation

type HsMultAnn pass = HsMultAnnOf (LHsType (NoGhcTc pass)) pass

data HsMultAnnOf mult pass

Denotes multiplicity annotations in the surface language. The mult type argument is usually `LHsType (NoGhcTc pass)`, but when the annotation is part of a type used in a term, it is `LHsExpr pass`. See Note [Types in terms].

Constructors

HsUnannotated !(XUnannotated mult pass)
a -> b or a → b or { nm :: a }

HsLinearAnn !(XLinearAnn mult pass)
a %1 -> b or a %1 → b, or a ⊸ b, or { nm %1 :: a }

HsExplicitMult !(XExplicitMult mult pass) !mult
a %m -> b or a %m → b or { nm %m :: a } (very much including `a %Many -> b`! This is how the programmer wrote it). It is stored as an HsType so as to preserve the syntax as written in the program.

XMultAnnOf !(XXMultAnnOf mult pass)

type family XUnannotated mult p

type family XLinearAnn mult p

type family XExplicitMult mult p

type family XXMultAnnOf mult p

data HsType pass

Haskell Type

Constructors

HsForAllTy

Fields

hst_xforall :: XForAllTy pass

hst_tele :: HsForAllTelescope pass

hst_body :: LHsType pass

HsQualTy

Fields

hst_xqual :: XQualTy pass

hst_ctxt :: LHsContext pass

hst_body :: LHsType pass

HsTyVar (XTyVar pass) PromotionFlag (LIdOccP pass)

HsAppTy (XAppTy pass) (LHsType pass) (LHsType pass)

HsAppKindTy (XAppKindTy pass) (LHsType pass) (LHsKind pass)

HsFunTy (XFunTy pass) (HsMultAnn pass) (LHsType pass) (LHsType pass)

HsListTy (XListTy pass) (LHsType pass)

HsTupleTy (XTupleTy pass) HsTupleSort [LHsType pass]

HsSumTy (XSumTy pass) [LHsType pass]

HsOpTy (XOpTy pass) PromotionFlag (LHsType pass) (LIdOccP pass) (LHsType pass)

HsParTy (XParTy pass) (LHsType pass)

HsIParamTy (XIParamTy pass) (XRec pass HsIPName) (LHsType pass)
(?x :: ty)

HsStarTy (XStarTy pass) Bool

HsKindSig (XKindSig pass) (LHsType pass) (LHsKind pass)
(ty :: kind)

HsSpliceTy (XSpliceTy pass) (HsUntypedSplice pass)

HsDocTy (XDocTy pass) (LHsType pass) (LHsDoc pass)

HsExplicitListTy (XExplicitListTy pass) PromotionFlag [LHsType pass]

HsExplicitTupleTy (XExplicitTupleTy pass) PromotionFlag [LHsType pass]

HsTyLit (XTyLit pass) (HsTyLit pass)

HsWildCardTy (XWildCardTy pass)

XHsType !(XXType pass)

type LHsType pass = XRec pass (HsType pass)

Located Haskell Type

type HsKind pass = HsType pass

Haskell Kind

type LHsKind pass = XRec pass (HsKind pass)

Located Haskell Kind

data HsBndrVis pass

Constructors

HsBndrRequired !(XBndrRequired pass)

HsBndrInvisible !(XBndrInvisible pass)

XBndrVis !(XXBndrVis pass)

type family XBndrRequired p

type family XBndrInvisible p

type family XXBndrVis p

data HsBndrVar pass

Constructors

HsBndrVar !(XBndrVar pass) !(LIdP pass)

HsBndrWildCard !(XBndrWildCard pass)

XBndrVar !(XXBndrVar pass)

type family XBndrVar p

type family XBndrWildCard p

type family XXBndrVar p

data HsBndrKind pass

Constructors

HsBndrKind !(XBndrKind pass) (LHsKind pass)

HsBndrNoKind !(XBndrNoKind pass)

XBndrKind !(XXBndrKind pass)

type family XBndrKind p

type family XBndrNoKind p

type family XXBndrKind p

isHsBndrInvisible :: HsBndrVis pass -> Bool

isHsBndrWildCard :: HsBndrVar pass -> Bool

data HsForAllTelescope pass

The type variable binders in an HsForAllTy. See also Note [Variable Specificity and Forall Visibility] in GHC.Tc.Gen.HsType.

Constructors

HsForAllVis
A visible forall (e.g., forall a -> {...}). These do not have any notion of specificity, so we use () as a placeholder value.

Fields

hsf_xvis :: XHsForAllVis pass

hsf_vis_bndrs :: [LHsTyVarBndr () pass]

HsForAllInvis
An invisible forall (e.g., forall a {b} c. {...}), where each binder has a Specificity.

Fields

hsf_xinvis :: XHsForAllInvis pass

hsf_invis_bndrs :: [LHsTyVarBndr Specificity pass]

XHsForAllTelescope !(XXHsForAllTelescope pass)

data HsTyVarBndr flag pass

Haskell Type Variable Binder See Note [Type variable binders]

Constructors

HsTvb

Fields

tvb_ext :: XTyVarBndr pass

tvb_flag :: flag

tvb_var :: HsBndrVar pass

tvb_kind :: HsBndrKind pass

XTyVarBndr !(XXTyVarBndr pass)

type LHsTyVarBndr flag pass = XRec pass (HsTyVarBndr flag pass)

Located Haskell Type Variable Binder

data LHsQTyVars pass

Located Haskell Quantified Type Variables

Constructors

HsQTvs

Fields

hsq_ext :: XHsQTvs pass

hsq_explicit :: [LHsTyVarBndr (HsBndrVis pass) pass]

XLHsQTyVars !(XXLHsQTyVars pass)

data HsOuterTyVarBndrs flag pass

The outermost type variables in a type that obeys the forall-or-nothing rule. See Note [forall-or-nothing rule].

Constructors

HsOuterImplicit
Implicit forall, e.g., f :: a -> b -> b

Fields

hso_ximplicit :: XHsOuterImplicit pass

HsOuterExplicit
Explicit forall, e.g., f :: forall a b. a -> b -> b

Fields

hso_xexplicit :: XHsOuterExplicit pass flag

hso_bndrs :: [LHsTyVarBndr flag (NoGhcTc pass)]

XHsOuterTyVarBndrs !(XXHsOuterTyVarBndrs pass)

type HsOuterFamEqnTyVarBndrs = HsOuterTyVarBndrs ()

Used for type-family instance equations, e.g.,

type instance forall a. F [a] = Tree a

The notion of specificity is irrelevant in type family equations, so we use () for the HsOuterTyVarBndrs flag.

type HsOuterSigTyVarBndrs = HsOuterTyVarBndrs Specificity

Used for signatures, e.g.,

f :: forall a {b}. blah

We use Specificity for the HsOuterTyVarBndrs flag to allow distinguishing between specified and inferred type variables.

data HsWildCardBndrs pass thing

Haskell Wildcard Binders

Constructors

HsWC

Fields

hswc_ext :: XHsWC pass thing

hswc_body :: thing

XHsWildCardBndrs !(XXHsWildCardBndrs pass thing)

data HsPatSigType pass

Types that can appear in pattern signatures, as well as the signatures for term-level binders in RULES. See Note [Pattern signature binders and scoping].

This is very similar to HsSigWcType, but with slightly different semantics: see Note [HsType binders]. See also Note [The wildcard story for types].

Constructors

HsPS

Fields

hsps_ext :: XHsPS pass

After renamer: HsPSRn

hsps_body :: LHsType pass

Main payload (the type itself)

XHsPatSigType !(XXHsPatSigType pass)

data HsSigType pass

A type signature that obeys the forall-or-nothing rule. In other words, an LHsType that uses an HsOuterSigTyVarBndrs to represent its outermost type variable quantification. See Note [Representing type signatures].

Constructors

HsSig

Fields

sig_ext :: XHsSig pass

sig_bndrs :: HsOuterSigTyVarBndrs pass

sig_body :: LHsType pass

XHsSigType !(XXHsSigType pass)

type LHsSigType pass = XRec pass (HsSigType pass)

Located Haskell Signature Type

type LHsSigWcType pass = HsWildCardBndrs pass (LHsSigType pass)

Located Haskell Signature Wildcard Type

type LHsWcType pass = HsWildCardBndrs pass (LHsType pass)

Located Haskell Wildcard Type

data HsTyPat pass

Constructors

HsTP

Fields

hstp_ext :: XHsTP pass

After renamer: HsTyPatRn

hstp_body :: LHsType pass

Main payload (the type itself)

XHsTyPat !(XXHsTyPat pass)

type LHsTyPat pass = XRec pass (HsTyPat pass)

data HsTupleSort

Haskell Tuple Sort

Constructors

HsUnboxedTuple

HsBoxedOrConstraintTuple

type HsContext pass = [LHsType pass]

Haskell Context

type LHsContext pass = XRec pass (HsContext pass)

Located Haskell Context

data HsTyLit pass

Haskell Type Literal

Constructors

HsNumTy (XNumTy pass) Integer

HsStrTy (XStrTy pass) FastString

HsCharTy (XCharTy pass) Char

XTyLit !(XXTyLit pass)

newtype HsIPName

These names are used early on to store the names of implicit parameters. They completely disappear after type-checking.

Constructors

HsIPName FastString

hsIPNameFS :: HsIPName -> FastString

data HsArg p tm ty

Arguments in an expression/type after splitting

Constructors

HsValArg !(XValArg p) tm

HsTypeArg !(XTypeArg p) ty

HsArgPar !(XArgPar p)

XArg !(XXArg p)

type family XValArg p

type family XTypeArg p

type family XArgPar p

type family XXArg p

type LHsTypeArg p = HsArg p (LHsType p) (LHsKind p)

data PromotionFlag

Is a TyCon a promoted data constructor or just a normal type constructor?

Constructors

NotPromoted

IsPromoted

isPromoted :: PromotionFlag -> Bool

data HsConDeclRecField pass

Constructor Declaration Record Field

Constructors

HsConDeclRecField

Fields

cdrf_ext :: XConDeclRecField pass

cdrf_names :: [LFieldOcc pass]

See Note [FieldOcc pass]

cdrf_spec :: HsConDeclField pass

XConDeclRecField !(XXConDeclRecField pass)

type LHsConDeclRecField pass = XRec pass (HsConDeclRecField pass)

Located Constructor Declaration Record Field

data HsConDetails arg rec

Describes the arguments to a data constructor. This is a common representation for several constructor-related concepts, including:

The arguments in a Haskell98-style constructor declaration (see HsConDeclH98Details in GHC.Hs.Decls).

The arguments in constructor patterns in case/function definitions (see HsConPatDetails in GHC.Hs.Pat).

The left-hand side arguments in a pattern synonym binding (see HsPatSynDetails in GHC.Hs.Binds).

One notable exception is the arguments in a GADT constructor, which uses a separate data type entirely (see HsConDeclGADTDetails in GHC.Hs.Decls). This is because GADT constructors cannot be declared with infix syntax, unlike the concepts above (#18844).

Constructors

PrefixCon [arg]

RecCon rec

InfixCon arg arg

data HsConDeclField pass

Constructor declaration field specification, see Note [HsConDeclField on pass]

Constructors

CDF

Fields

cdf_ext :: XConDeclField pass

Extension point

cdf_unpack :: SrcUnpackedness

UNPACK pragma if any E.g. data T = MkT {-# UNPACK #-} Int or data T where MtT :: {-# UNPACK #-} Int -> T

cdf_bang :: SrcStrictness

User-specified strictness, if any E.g. data T a = MkT !a or data T a where MtT :: !a -> T a

cdf_multiplicity :: HsMultAnn pass

User-specified multiplicity, if any E.g. data T a = MkT { t %Many :: a } or data T a where MtT :: a %1 -> T a

cdf_type :: LHsType pass

The type of the field

cdf_doc :: Maybe (LHsDoc pass)

Documentation for the field F.e. this very piece of documentation

data FieldOcc pass

Field Occurrence

Represents an *occurrence* of a field. This may or may not be a binding occurrence (e.g. this type is used in HsConDeclRecField and RecordPatSynField which bind their fields, but also in HsRecField for record construction and patterns, which do not).

We store both the RdrName the user originally wrote, and after the renamer we use the extension field to store the selector function. See note [FieldOcc pass]

There is a wrinkle in that update field occurances are sometimes ambiguous during the rename stage. See note [Ambiguous FieldOcc in record updates] to see how we currently handle this.

Constructors

FieldOcc

Fields

foExt :: XCFieldOcc pass

foLabel :: LIdP pass

XFieldOcc !(XXFieldOcc pass)

type LFieldOcc pass = XRec pass (FieldOcc pass)

Located Field Occurrence

mapHsOuterImplicit :: (XHsOuterImplicit pass -> XHsOuterImplicit pass) -> HsOuterTyVarBndrs flag pass -> HsOuterTyVarBndrs flag pass

hsQTvExplicit :: LHsQTyVars pass -> [LHsTyVarBndr (HsBndrVis pass) pass]

isHsKindedTyVar :: HsTyVarBndr flag pass -> Bool

Does this HsTyVarBndr come with an explicit kind annotation?