LinkerTypes

Documentation

newtype DynLinker

Constructors

DynLinker

Fields

dl_mpls :: MVar (Maybe PersistentLinkerState)

data PersistentLinkerState

Constructors

PersistentLinkerState

Fields

closure_env :: ClosureEnv

itbl_env :: !ItblEnv

bcos_loaded :: ![Linkable]

objs_loaded :: ![Linkable]

pkgs_loaded :: ![LinkerUnitId]

temp_sos :: ![(FilePath, String)]

type LinkerUnitId = InstalledUnitId

data Linkable

Information we can use to dynamically link modules into the compiler

Constructors

LM

Fields

linkableTime :: UTCTime

Time at which this linkable was built (i.e. when the bytecodes were produced, or the mod date on the files)

linkableModule :: Module

The linkable module itself

linkableUnlinked :: [Unlinked]

Those files and chunks of code we have yet to link.

INVARIANT: A valid linkable always has at least one Unlinked item. If this list is empty, the Linkable represents a fake linkable, which is generated in HscNothing mode to avoid recompiling modules.

ToDo: Do items get removed from this list when they get linked?

data Unlinked

Objects which have yet to be linked by the compiler

Constructors

DotO FilePath
An object file (.o)

DotA FilePath
Static archive file (.a)

DotDLL FilePath
Dynamically linked library file (.so, .dll, .dylib)

BCOs CompiledByteCode [SptEntry]
A byte-code object, lives only in memory. Also carries some static pointer table entries which should be loaded along with the BCOs. See Note [Grant plan for static forms] in StaticPtrTable.

data SptEntry

An entry to be inserted into a module's static pointer table. See Note [Grand plan for static forms] in StaticPtrTable.

Constructors

SptEntry Id Fingerprint