ListT

Description

A backtracking, logic programming monad.

Adapted from the paper /Backtracking, Interleaving, and Terminating Monad Transformers/, by Oleg Kiselyov, Chung-chieh Shan, Daniel P. Friedman, Amr Sabry (http://www.cs.rutgers.edu/~ccshan/logicprog/ListT-icfp2005.pdf).

Documentation

newtype ListT m a

A monad transformer for performing backtracking computations layered over another monad m

Constructors

ListT

Fields

unListT :: forall r. (a -> m r -> m r) -> m r -> m r

runListT :: ListT m a -> (a -> m r -> m r) -> m r -> m r

Runs a ListT computation with the specified initial success and failure continuations.

select :: Monad m => [a] -> ListT m a

fold :: ListT m a -> (a -> m r -> m r) -> m r -> m r