Llvm.PpLlvm

Description

Pretty print LLVM IR Code.

Top level LLVM objects.

ppLlvmModule :: LlvmModule -> SDoc

Print out a whole LLVM module.

ppLlvmComments :: [LMString] -> SDoc

Print out a multi-line comment, can be inside a function or on its own

ppLlvmComment :: LMString -> SDoc

Print out a comment, can be inside a function or on its own

ppLlvmGlobals :: [LMGlobal] -> SDoc

Print out a list of global mutable variable definitions

ppLlvmGlobal :: LMGlobal -> SDoc

Print out a global mutable variable definition

ppLlvmAliases :: [LlvmAlias] -> SDoc

Print out a list of LLVM type aliases.

ppLlvmAlias :: LlvmAlias -> SDoc

Print out an LLVM type alias.

ppLlvmMetas :: [MetaDecl] -> SDoc

Print out a list of LLVM metadata.

ppLlvmMeta :: MetaDecl -> SDoc

Print out an LLVM metadata definition.

ppLlvmFunctionDecls :: LlvmFunctionDecls -> SDoc

Print out a list of function declaration.

ppLlvmFunctionDecl :: LlvmFunctionDecl -> SDoc

Print out a function declaration. Declarations define the function type but don't define the actual body of the function.

ppLlvmFunctions :: LlvmFunctions -> SDoc

Print out a list of function definitions.

ppLlvmFunction :: LlvmFunction -> SDoc

Print out a function definition.