LlvmCodeGen.Base

Description

Base LLVM Code Generation module

Contains functions useful through out the code generator.

Documentation

type LlvmCmmDecl = GenCmmDecl [LlvmData] (Maybe CmmStatics) (ListGraph LlvmStatement)

type LlvmBasicBlock = GenBasicBlock LlvmStatement

type LiveGlobalRegs = [GlobalReg]

Global registers live on proc entry

type LlvmUnresData = (CLabel, Section, LlvmType, [UnresStatic])

Unresolved code. Of the form: (data label, data type, unresolved data)

type LlvmData = ([LMGlobal], [LlvmType])

Top level LLVM Data (globals and type aliases)

type UnresLabel = CmmLit

An unresolved Label.

Labels are unresolved when we haven't yet determined if they are defined in the module we are currently compiling, or an external one.

type UnresStatic = Either UnresLabel LlvmStatic

type LlvmVersion = (Int, Int)

LLVM Version Number

supportedLlvmVersion :: LlvmVersion

The LLVM Version that is currently supported.

llvmVersionStr :: LlvmVersion -> String

data LlvmM a

The Llvm monad. Wraps LlvmEnv state as well as the IO monad

runLlvm :: DynFlags -> LlvmVersion -> BufHandle -> UniqSupply -> LlvmM () -> IO ()

Get initial Llvm environment.

liftStream :: Stream IO a x -> Stream LlvmM a x

Lift a stream into the LlvmM monad

withClearVars :: LlvmM a -> LlvmM a

Clear variables from the environment for a subcomputation

varLookup :: Uniquable key => key -> LlvmM (Maybe LlvmType)

Lookup variables or functions in the environment.

varInsert :: Uniquable key => key -> LlvmType -> LlvmM ()

Insert variables or functions into the environment.

markStackReg :: GlobalReg -> LlvmM ()

Set a register as allocated on the stack

checkStackReg :: GlobalReg -> LlvmM Bool

Check whether a register is allocated on the stack

funLookup :: Uniquable key => key -> LlvmM (Maybe LlvmType)

Lookup variables or functions in the environment.

funInsert :: Uniquable key => key -> LlvmType -> LlvmM ()

Insert variables or functions into the environment.

getLlvmVer :: LlvmM LlvmVersion

Get the LLVM version we are generating code for

getDynFlags :: HasDynFlags m => m DynFlags

getDynFlag :: (DynFlags -> a) -> LlvmM a

Get the platform we are generating code for

getLlvmPlatform :: LlvmM Platform

Get the platform we are generating code for

dumpIfSetLlvm :: DumpFlag -> String -> SDoc -> LlvmM ()

Dumps the document if the corresponding flag has been set by the user

renderLlvm :: SDoc -> LlvmM ()

Prints the given contents to the output handle

markUsedVar :: LlvmVar -> LlvmM ()

Marks a variable as "used"

getUsedVars :: LlvmM [LlvmVar]

Return all variables marked as "used" so far

ghcInternalFunctions :: LlvmM ()

Here we pre-initialise some functions that are used internally by GHC so as to make sure they have the most general type in the case that user code also uses these functions but with a different type than GHC internally. (Main offender is treating return type as void instead of 'void *'). Fixes trac #5486.

getMetaUniqueId :: LlvmM MetaId

Allocate a new global unnamed metadata identifier

setUniqMeta :: Unique -> MetaId -> LlvmM ()

Sets metadata node for a given unique

getUniqMeta :: Unique -> LlvmM (Maybe MetaId)

Gets metadata node for given unique

cmmToLlvmType :: CmmType -> LlvmType

Translate a basic CmmType to an LlvmType.

widthToLlvmFloat :: Width -> LlvmType

Translate a Cmm Float Width to a LlvmType.

widthToLlvmInt :: Width -> LlvmType

Translate a Cmm Bit Width to a LlvmType.

llvmFunTy :: LiveGlobalRegs -> LlvmM LlvmType

Llvm Function type for Cmm function

llvmFunSig :: LiveGlobalRegs -> CLabel -> LlvmLinkageType -> LlvmM LlvmFunctionDecl

Llvm Function signature

llvmFunArgs :: DynFlags -> LiveGlobalRegs -> [LlvmVar]

A Function's arguments

llvmStdFunAttrs :: [LlvmFuncAttr]

Llvm standard fun attributes

llvmFunAlign :: DynFlags -> LMAlign

Alignment to use for functions

llvmInfAlign :: DynFlags -> LMAlign

Alignment to use for into tables

llvmPtrBits :: DynFlags -> Int

Pointer width

tysToParams :: [LlvmType] -> [LlvmParameter]

Convert a list of types to a list of function parameters (each with no parameter attributes)

llvmFunSection :: DynFlags -> LMString -> LMSection

Section to use for a function

strCLabel_llvm :: CLabel -> LlvmM LMString

Pretty print a CLabel.

strDisplayName_llvm :: CLabel -> LlvmM LMString

strProcedureName_llvm :: CLabel -> LlvmM LMString

getGlobalPtr :: LMString -> LlvmM LlvmVar

Create/get a pointer to a global value. Might return an alias if the value in question hasn't been defined yet. We especially make no guarantees on the type of the returned pointer.

generateExternDecls :: LlvmM ([LMGlobal], [LlvmType])

Generate definitions for aliases forward-referenced by getGlobalPtr.

Must be called at a point where we are sure that no new global definitions will be generated anymore!

aliasify :: LMGlobal -> LlvmM [LMGlobal]

Here we take a global variable definition, rename it with a $def suffix, and generate the appropriate alias.