LlvmCodeGen.Data

Description

Handle conversion of CmmData to LLVM code.

Documentation

genLlvmData :: LlvmEnv -> (Section, CmmStatics) -> LlvmUnresData

Pass a CmmStatic section to an equivalent Llvm code. Can't complete this completely though as we need to pass all CmmStatic sections before all references can be resolved. This last step is done by resolveLlvmData.

resolveLlvmDatas :: LlvmEnv -> [LlvmUnresData] -> (LlvmEnv, [LlvmData])

resolveLlvmData :: LlvmEnv -> LlvmUnresData -> (LlvmEnv, LlvmData)

Fix up CLabel references now that we should have passed all CmmData.