MkIface

Description

Module for constructing ModIface values (interface files), writing them to disk and comparing two versions to see if recompilation is required.

Documentation

mkIface :: HscEnv -> Maybe Fingerprint -> ModDetails -> ModGuts -> IO (ModIface, Bool)

mkIfaceTc :: HscEnv -> Maybe Fingerprint -> SafeHaskellMode -> ModDetails -> TcGblEnv -> IO (ModIface, Bool)

make an interface from the results of typechecking only. Useful for non-optimising compilation, or where we aren't generating any object code at all (HscNothing).

writeIfaceFile :: DynFlags -> FilePath -> ModIface -> IO ()

checkOldIface :: HscEnv -> ModSummary -> SourceModified -> Maybe ModIface -> IO (RecompileRequired, Maybe ModIface)

Top level function to check if the version of an old interface file is equivalent to the current source file the user asked us to compile. If the same, we can avoid recompilation. We return a tuple where the first element is a bool saying if we should recompile the object file and the second is maybe the interface file, where Nothng means to rebuild the interface file not use the exisitng one.

data RecompileRequired

Constructors

UpToDate
everything is up to date, recompilation is not required

MustCompile
The .hs file has been touched, or the .o/.hi file does not exist

RecompBecause String
The .o/.hi files are up to date, but something else has changed to force recompilation; the String says what (one-line summary)

recompileRequired :: RecompileRequired -> Bool

mkIfaceExports :: [AvailInfo] -> [IfaceExport]

tyThingToIfaceDecl :: TyThing -> IfaceDecl