NameCache

Description

The Name Cache

Documentation

lookupOrigNameCache :: OrigNameCache -> Module -> OccName -> Maybe Name

extendOrigNameCache :: OrigNameCache -> Name -> OrigNameCache

extendNameCache :: OrigNameCache -> Module -> OccName -> Name -> OrigNameCache

initNameCache :: UniqSupply -> [Name] -> NameCache

Return a function to atomically update the name cache.

data NameCache

The NameCache makes sure that there is just one Unique assigned for each original name; i.e. (module-name, occ-name) pair and provides something of a lookup mechanism for those names.

Constructors

NameCache

Fields

nsUniqs :: !UniqSupply

Supply of uniques

nsNames :: !OrigNameCache

Ensures that one original name gets one unique

type OrigNameCache = ModuleEnv (OccEnv Name)

Per-module cache of original OccNames given Names