OccName

Description

GHC uses several kinds of name internally:

OccName represents names as strings with just a little more information: the "namespace" that the name came from, e.g. the namespace of value, type constructors or data constructors

RdrName: see RdrName

Name: see Name

Id: see Id

Var: see Var

The NameSpace type

data NameSpace

nameSpacesRelated :: NameSpace -> NameSpace -> Bool

Construction

There are two forms of data constructor:

Source data constructors

The data constructors mentioned in Haskell source code

Real data constructors

The data constructors of the representation type, which may not be the same as the source type

For example:

data T = T !(Int, Int)

The source datacon has type (Int, Int) -> T The real datacon has type Int -> Int -> T

GHC chooses a representation based on the strictness etc.

tcName :: NameSpace

clsName :: NameSpace

tcClsName :: NameSpace

dataName :: NameSpace

varName :: NameSpace

tvName :: NameSpace

srcDataName :: NameSpace

Pretty Printing

pprNameSpace :: NameSpace -> SDoc

pprNonVarNameSpace :: NameSpace -> SDoc

pprNameSpaceBrief :: NameSpace -> SDoc

The OccName type

data OccName

Occurrence Name

In this context that means: "classified (i.e. as a type name, value name, etc) but not qualified and not yet resolved"

pprOccName :: OccName -> SDoc

Construction

mkOccName :: NameSpace -> String -> OccName

mkOccNameFS :: NameSpace -> FastString -> OccName

mkVarOcc :: String -> OccName

mkVarOccFS :: FastString -> OccName

mkDataOcc :: String -> OccName

mkDataOccFS :: FastString -> OccName

mkTyVarOcc :: String -> OccName

mkTyVarOccFS :: FastString -> OccName

mkTcOcc :: String -> OccName

mkTcOccFS :: FastString -> OccName

mkClsOcc :: String -> OccName

mkClsOccFS :: FastString -> OccName

mkDFunOcc

Arguments

:: String
Typically the class and type glommed together e.g. OrdMaybe. Only used in debug mode, for extra clarity

-> Bool
Is this a hs-boot instance DFun?

-> OccSet
avoid these Occs

-> OccName
E.g. $f3OrdMaybe

setOccNameSpace :: NameSpace -> OccName -> OccName

demoteOccName :: OccName -> Maybe OccName

class HasOccName name where

Other names in the compiler add additional information to an OccName. This class provides a consistent way to access the underlying OccName.

Minimal complete definition

occName

Methods

occName :: name -> OccName

Derived OccNames

isDerivedOccName :: OccName -> Bool

Test for definitions internally generated by GHC. This predicte is used to suppress printing of internal definitions in some debug prints

mkDataConWrapperOcc :: OccName -> OccName

mkWorkerOcc :: OccName -> OccName

mkMatcherOcc :: OccName -> OccName

mkBuilderOcc :: OccName -> OccName

mkDefaultMethodOcc :: OccName -> OccName

isDefaultMethodOcc :: OccName -> Bool

isTypeableBindOcc :: OccName -> Bool

Is an OccName one of a Typeable TyCon or Module binding? This is needed as these bindings are renamed differently. See Note [Grand plan for Typeable] in TcTypeable.

mkNewTyCoOcc :: OccName -> OccName

mkClassOpAuxOcc :: OccName -> OccName

mkCon2TagOcc :: OccName -> OccName

mkTag2ConOcc :: OccName -> OccName

mkMaxTagOcc :: OccName -> OccName

mkClassDataConOcc :: OccName -> OccName

mkDictOcc :: OccName -> OccName

mkIPOcc :: OccName -> OccName

mkSpecOcc :: OccName -> OccName

mkForeignExportOcc :: OccName -> OccName

mkRepEqOcc :: OccName -> OccName

mkGenR :: OccName -> OccName

mkGen1R :: OccName -> OccName

mkDataTOcc

Arguments

:: OccName
TyCon or data con string

-> OccSet
avoid these Occs

-> OccName
E.g. $f3OrdMaybe data T = MkT ... deriving( Data ) needs definitions for $tT :: Data.Generics.Basics.DataType $cMkT :: Data.Generics.Basics.Constr

mkDataCOcc

Arguments

:: OccName
TyCon or data con string

-> OccSet
avoid these Occs

-> OccName
E.g. $f3OrdMaybe data T = MkT ... deriving( Data ) needs definitions for $tT :: Data.Generics.Basics.DataType $cMkT :: Data.Generics.Basics.Constr

mkDataConWorkerOcc :: OccName -> OccName

mkSuperDictSelOcc

Arguments

:: Int
Index of superclass, e.g. 3

-> OccName
Class, e.g. Ord

-> OccName
Derived Occname, e.g. $p3Ord

mkSuperDictAuxOcc :: Int -> OccName -> OccName

mkLocalOcc

Arguments

:: Unique
Unique to combine with the OccName

-> OccName
Local name, e.g. sat

-> OccName
Nice unique version, e.g. $L23sat

mkMethodOcc :: OccName -> OccName

mkInstTyTcOcc

Arguments

:: String
Family name, e.g. Map

-> OccSet
avoid these Occs

-> OccName
R:Map

Derive a name for the representation type constructor of a data/newtype instance.

mkInstTyCoOcc :: OccName -> OccName

mkEqPredCoOcc :: OccName -> OccName

mkVectOcc :: Maybe String -> OccName -> OccName

mkVectTyConOcc :: Maybe String -> OccName -> OccName

mkVectDataConOcc :: Maybe String -> OccName -> OccName

mkVectIsoOcc :: Maybe String -> OccName -> OccName

mkPDataTyConOcc :: Maybe String -> OccName -> OccName

mkPDataDataConOcc :: Maybe String -> OccName -> OccName

mkPDatasTyConOcc :: Maybe String -> OccName -> OccName

mkPDatasDataConOcc :: Maybe String -> OccName -> OccName

mkPReprTyConOcc :: Maybe String -> OccName -> OccName

mkPADFunOcc :: Maybe String -> OccName -> OccName

mkRecFldSelOcc :: String -> OccName

mkTyConRepOcc :: OccName -> OccName

Deconstruction

occNameFS :: OccName -> FastString

occNameString :: OccName -> String

occNameSpace :: OccName -> NameSpace

isVarOcc :: OccName -> Bool

isTvOcc :: OccName -> Bool

isTcOcc :: OccName -> Bool

isDataOcc :: OccName -> Bool

isDataSymOcc :: OccName -> Bool

Test if the OccName is a data constructor that starts with a symbol (e.g. :, or [])

isSymOcc :: OccName -> Bool

Test if the OccName is that for any operator (whether it is a data constructor or variable or whatever)

isValOcc :: OccName -> Bool

Value OccNamess are those that are either in the variable or data constructor namespaces

parenSymOcc :: OccName -> SDoc -> SDoc

Wrap parens around an operator

startsWithUnderscore :: OccName -> Bool

Haskell 98 encourages compilers to suppress warnings about unsed names in a pattern if they start with _: this implements that test

isTcClsNameSpace :: NameSpace -> Bool

isTvNameSpace :: NameSpace -> Bool

isDataConNameSpace :: NameSpace -> Bool

isVarNameSpace :: NameSpace -> Bool

isValNameSpace :: NameSpace -> Bool

The OccEnv type

data OccEnv a

emptyOccEnv :: OccEnv a

unitOccEnv :: OccName -> a -> OccEnv a

extendOccEnv :: OccEnv a -> OccName -> a -> OccEnv a

mapOccEnv :: (a -> b) -> OccEnv a -> OccEnv b

lookupOccEnv :: OccEnv a -> OccName -> Maybe a

mkOccEnv :: [(OccName, a)] -> OccEnv a

mkOccEnv_C :: (a -> a -> a) -> [(OccName, a)] -> OccEnv a

extendOccEnvList :: OccEnv a -> [(OccName, a)] -> OccEnv a

elemOccEnv :: OccName -> OccEnv a -> Bool

occEnvElts :: OccEnv a -> [a]

foldOccEnv :: (a -> b -> b) -> b -> OccEnv a -> b

plusOccEnv :: OccEnv a -> OccEnv a -> OccEnv a

plusOccEnv_C :: (a -> a -> a) -> OccEnv a -> OccEnv a -> OccEnv a

extendOccEnv_C :: (a -> a -> a) -> OccEnv a -> OccName -> a -> OccEnv a

extendOccEnv_Acc :: (a -> b -> b) -> (a -> b) -> OccEnv b -> OccName -> a -> OccEnv b

filterOccEnv :: (elt -> Bool) -> OccEnv elt -> OccEnv elt

delListFromOccEnv :: OccEnv a -> [OccName] -> OccEnv a

delFromOccEnv :: OccEnv a -> OccName -> OccEnv a

alterOccEnv :: (Maybe elt -> Maybe elt) -> OccEnv elt -> OccName -> OccEnv elt

pprOccEnv :: (a -> SDoc) -> OccEnv a -> SDoc

The OccSet type

type OccSet = UniqSet OccName

emptyOccSet :: OccSet

unitOccSet :: OccName -> OccSet

mkOccSet :: [OccName] -> OccSet

extendOccSet :: OccSet -> OccName -> OccSet

extendOccSetList :: OccSet -> [OccName] -> OccSet

unionOccSets :: OccSet -> OccSet -> OccSet

unionManyOccSets :: [OccSet] -> OccSet

minusOccSet :: OccSet -> OccSet -> OccSet

elemOccSet :: OccName -> OccSet -> Bool

isEmptyOccSet :: OccSet -> Bool

intersectOccSet :: OccSet -> OccSet -> OccSet

intersectsOccSet :: OccSet -> OccSet -> Bool

filterOccSet :: (OccName -> Bool) -> OccSet -> OccSet

Tidying up

type TidyOccEnv = UniqFM Int

emptyTidyOccEnv :: TidyOccEnv

initTidyOccEnv :: [OccName] -> TidyOccEnv

tidyOccName :: TidyOccEnv -> OccName -> (TidyOccEnv, OccName)

tidyOccNames :: TidyOccEnv -> [OccName] -> (TidyOccEnv, [OccName])

avoidClashesOccEnv :: TidyOccEnv -> [OccName] -> TidyOccEnv

type FastStringEnv a = UniqFM a

A non-deterministic set of FastStrings. See Note [Deterministic UniqFM] in UniqDFM for explanation why it's not deterministic and why it matters. Use DFastStringEnv if the set eventually gets converted into a list or folded over in a way where the order changes the generated code.

emptyFsEnv :: FastStringEnv a

lookupFsEnv :: FastStringEnv a -> FastString -> Maybe a

extendFsEnv :: FastStringEnv a -> FastString -> a -> FastStringEnv a

mkFsEnv :: [(FastString, a)] -> FastStringEnv a