OldCmm

Documentation

type CmmGroup = GenCmmGroup CmmStatics CmmInfo (ListGraph CmmStmt)

Cmm with the info table as a data type

type GenCmmGroup d h g = [GenCmmDecl d h g]

type RawCmmGroup = GenCmmGroup CmmStatics (Maybe CmmStatics) (ListGraph CmmStmt)

Cmm with the info tables converted to a list of CmmStatic along with the info table label. If we are building without tables-next-to-code there will be no statics

INVARIANT: if there is an info table, it has at least one CmmStatic

type CmmDecl = GenCmmDecl CmmStatics CmmInfo (ListGraph CmmStmt)

type RawCmmDecl = GenCmmDecl CmmStatics (Maybe CmmStatics) (ListGraph CmmStmt)

newtype ListGraph i

A control-flow graph represented as a list of extended basic blocks.

Constructors

ListGraph [GenBasicBlock i]
Code, may be empty. The first block is the entry point. The order is otherwise initially unimportant, but at some point the code gen will fix the order.

data CmmInfo

Constructors

CmmInfo (Maybe BlockId) (Maybe UpdateFrame) CmmInfoTable

data UpdateFrame

A frame that is to be pushed before entry to the function. Used to handle update frames.

Constructors

UpdateFrame CmmExpr [CmmExpr]

data CmmInfoTable

Info table as a haskell data type

Constructors

CmmInfoTable

Fields

cit_lbl :: CLabel

cit_rep :: SMRep

cit_prof :: ProfilingInfo

cit_srt :: C_SRT

CmmNonInfoTable

data ClosureTypeInfo

Constructors

Constr ConstrTag ConstrDescription

Fun FunArity ArgDescr

Thunk

ThunkSelector SelectorOffset

BlackHole

data CmmStatic

Constructors

CmmStaticLit CmmLit

CmmUninitialised Int

CmmString [Word8]

data CmmStatics

Constructors

Statics CLabel [CmmStatic]

type CmmFormal = LocalReg

type CmmActual = CmmExpr

cmmMapGraph :: (g -> g') -> GenCmmGroup d h g -> GenCmmGroup d h g'

cmmTopMapGraph :: (g -> g') -> GenCmmDecl d h g -> GenCmmDecl d h g'

data GenBasicBlock i

Constructors

BasicBlock BlockId [i]

type CmmBasicBlock = GenBasicBlock CmmStmt

blockId :: GenBasicBlock i -> BlockId

blockStmts :: GenBasicBlock i -> [i]

mapBlockStmts :: (i -> i') -> GenBasicBlock i -> GenBasicBlock i'

data CmmStmt

Constructors

CmmNop

CmmComment FastString

CmmAssign CmmReg CmmExpr

CmmStore CmmExpr CmmExpr

CmmCall CmmCallTarget [HintedCmmFormal] [HintedCmmActual] CmmReturnInfo

CmmBranch BlockId

CmmCondBranch CmmExpr BlockId

CmmSwitch CmmExpr [Maybe BlockId]

CmmJump CmmExpr [HintedCmmActual]

CmmReturn [HintedCmmActual]

data CmmReturnInfo

Constructors

CmmMayReturn

CmmNeverReturns

data CmmHinted a

Constructors

CmmHinted

Fields

hintlessCmm :: a

cmmHint :: ForeignHint

type HintedCmmFormal = CmmHinted CmmFormal

type HintedCmmActual = CmmHinted CmmActual

data CmmSafety

Constructors

CmmUnsafe

CmmSafe C_SRT

CmmInterruptible

data CmmCallTarget

Constructors

CmmCallee CmmExpr CCallConv

CmmPrim CallishMachOp

data GenCmmDecl d h g

A top-level chunk, abstracted over the type of the contents of the basic blocks (Cmm or instructions are the likely instantiations).

Constructors

CmmProc h CLabel g

CmmData Section d

data ForeignHint

Constructors

NoHint

AddrHint

SignedHint

module CmmExpr

data Section

Constructors

Text

Data

ReadOnlyData

RelocatableReadOnlyData

UninitialisedData

ReadOnlyData16

OtherSection String

data ProfilingInfo

Constructors

NoProfilingInfo

ProfilingInfo [Word8] [Word8]

data C_SRT

Constructors

NoC_SRT

C_SRT !CLabel !WordOff !StgHalfWord