PackageConfig

Description

Package configuration information: essentially the interface to Cabal, with some utilities

(c) The University of Glasgow, 2004

Documentation

Mostly the compiler deals in terms of UnitIds, which are md5 hashes of a package ID, keys of its dependencies, and Cabal flags. You're expected to pass in the unit id in the -this-unit-id flag. However, for wired-in packages like base & rts, we don't necessarily know what the version is, so these are handled specially; see .

UnitId

packageConfigId :: PackageConfig -> UnitId

expandedPackageConfigId :: PackageConfig -> UnitId

definitePackageConfigId :: PackageConfig -> Maybe DefUnitId

installedPackageConfigId :: PackageConfig -> InstalledUnitId

Get the GHC UnitId right out of a Cabalish PackageConfig

The PackageConfig type: information about a package

type PackageConfig = InstalledPackageInfo ComponentId SourcePackageId PackageName InstalledUnitId UnitId ModuleName Module

data InstalledPackageInfo compid srcpkgid srcpkgname instunitid unitid modulename mod :: * -> * -> * -> * -> * -> * -> * -> *

This is a subset of Cabal's InstalledPackageInfo, with just the bits that GHC is interested in. See Cabal's documentation for a more detailed description of all of the fields.

Constructors

InstalledPackageInfo

Fields

unitId :: instunitid

componentId :: compid

instantiatedWith :: [(modulename, mod)]

sourcePackageId :: srcpkgid

packageName :: srcpkgname

packageVersion :: Version

sourceLibName :: Maybe srcpkgname

abiHash :: String

depends :: [instunitid]

abiDepends :: [(instunitid, String)]

Like depends, but each dependency is annotated with the ABI hash we expect the dependency to respect.

importDirs :: [FilePath]

hsLibraries :: [String]

extraLibraries :: [String]

extraGHCiLibraries :: [String]

libraryDirs :: [FilePath]

libraryDynDirs :: [FilePath]

frameworks :: [String]

frameworkDirs :: [FilePath]

ldOptions :: [String]

ccOptions :: [String]

includes :: [String]

includeDirs :: [FilePath]

haddockInterfaces :: [FilePath]

haddockHTMLs :: [FilePath]

exposedModules :: [(modulename, Maybe mod)]

hiddenModules :: [modulename]

indefinite :: Bool

exposed :: Bool

trusted :: Bool

newtype ComponentId

A ComponentId consists of the package name, package version, component ID, the transitive dependencies of the component, and other information to uniquely identify the source code and build configuration of a component.

This used to be known as an InstalledPackageId, but a package can contain multiple components and a ComponentId uniquely identifies a component within a package. When a package only has one component, the ComponentId coincides with the InstalledPackageId

Constructors

ComponentId FastString

newtype SourcePackageId

Constructors

SourcePackageId FastString

newtype PackageName

Constructors

PackageName FastString

data Version :: *

A Version represents the version of a software entity.

An instance of Eq is provided, which implements exact equality modulo reordering of the tags in the versionTags field.

An instance of Ord is also provided, which gives lexicographic ordering on the versionBranch fields (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.). This is expected to be sufficient for many uses, but note that you may need to use a more specific ordering for your versioning scheme. For example, some versioning schemes may include pre-releases which have tags "pre1", "pre2", and so on, and these would need to be taken into account when determining ordering. In some cases, date ordering may be more appropriate, so the application would have to look for date tags in the versionTags field and compare those. The bottom line is, don't always assume that compare and other Ord operations are the right thing for every Version.

Similarly, concrete representations of versions may differ. One possible concrete representation is provided (see showVersion and parseVersion), but depending on the application a different concrete representation may be more appropriate.

Constructors

Version

Fields

versionBranch :: [Int]

The numeric branch for this version. This reflects the fact that most software versions are tree-structured; there is a main trunk which is tagged with versions at various points (1,2,3...), and the first branch off the trunk after version 3 is 3.1, the second branch off the trunk after version 3 is 3.2, and so on. The tree can be branched arbitrarily, just by adding more digits.

We represent the branch as a list of Int, so version 3.2.1 becomes [3,2,1]. Lexicographic ordering (i.e. the default instance of Ord for [Int]) gives the natural ordering of branches.

versionTags :: [String]

A version can be tagged with an arbitrary list of strings. The interpretation of the list of tags is entirely dependent on the entity that this version applies to.

defaultPackageConfig :: PackageConfig

sourcePackageIdString :: PackageConfig -> String

packageNameString :: PackageConfig -> String

pprPackageConfig :: PackageConfig -> SDoc