PlaceHolder

Documentation

data PlaceHolder

used as place holder in PostTc and PostRn values

Constructors

PlaceHolder

type family PostTc id ty

Types that are not defined until after type checking

type family PostRn id ty

Types that are not defined until after renaming

placeHolderKind :: PlaceHolder

placeHolderFixity :: PlaceHolder

placeHolderType :: PlaceHolder

placeHolderTypeTc :: Type

placeHolderNames :: PlaceHolder

placeHolderNamesTc :: NameSet

placeHolderHsWrapper :: PlaceHolder

type DataId id = (DataIdPost id, DataIdPost (NameOrRdrName id))

type DataIdPost id = (Data id, Data (PostRn id NameSet), Data (PostRn id Fixity), Data (PostRn id Bool), Data (PostRn id Name), Data (PostRn id (Located Name)), Data (PostRn id [Name]), Data (PostRn id id), Data (PostTc id Type), Data (PostTc id Coercion), Data (PostTc id id), Data (PostTc id [Type]), Data (PostTc id ConLike), Data (PostTc id [ConLike]), Data (PostTc id HsWrapper), Data (PostTc id [FieldLabel]))

type family NameOrRdrName id where ...

Follow the id, but never beyond Name. This is used in a HsMatchContext, for printing messages related to a Match

Equations

NameOrRdrName Id = Name

NameOrRdrName Name = Name

NameOrRdrName RdrName = RdrName

type OutputableBndrId id = (OutputableBndr id, OutputableBndr (NameOrRdrName id))

Constraint type to bundle up the requirement for OutputableBndr on both the id and the NameOrRdrName type for it