Platform

Description

A description of the platform we're compiling for.

Documentation

data Platform

Contains enough information for the native code generator to emit code for this platform.

Constructors

Platform

Fields

platformArch :: Arch

platformOS :: OS

platformWordSize :: !Int

platformUnregisterised :: Bool

platformHasGnuNonexecStack :: Bool

platformHasIdentDirective :: Bool

platformHasSubsectionsViaSymbols :: Bool

platformIsCrossCompiling :: Bool

data Arch

Architectures that the native code generator knows about. TODO: It might be nice to extend these constructors with information about what instruction set extensions an architecture might support.

Constructors

ArchUnknown

ArchX86

ArchX86_64

ArchPPC

ArchPPC_64

Fields

ppc_64ABI :: PPC_64ABI

ArchSPARC

ArchSPARC64

ArchARM

Fields

armISA :: ArmISA

armISAExt :: [ArmISAExt]

armABI :: ArmABI

ArchARM64

ArchAlpha

ArchMipseb

ArchMipsel

ArchJavaScript

data OS

Operating systems that the native code generator knows about. Having OSUnknown should produce a sensible default, but no promises.

Constructors

OSUnknown

OSLinux

OSDarwin

OSSolaris2

OSMinGW32

OSFreeBSD

OSDragonFly

OSOpenBSD

OSNetBSD

OSKFreeBSD

OSHaiku

OSQNXNTO

OSAIX

OSHurd

data ArmISA

ARM Instruction Set Architecture, Extensions and ABI

Constructors

ARMv5

ARMv6

ARMv7

data ArmISAExt

Constructors

VFPv2

VFPv3

VFPv3D16

NEON

IWMMX2

data ArmABI

Constructors

SOFT

SOFTFP

HARD

data PPC_64ABI

PowerPC 64-bit ABI

Constructors

ELF_V1

ELF_V2

target32Bit :: Platform -> Bool

This predicate tells us whether the platform is 32-bit.

isARM :: Arch -> Bool

osElfTarget :: OS -> Bool

This predicate tells us whether the OS supports ELF-like shared libraries.

osMachOTarget :: OS -> Bool

This predicate tells us whether the OS support Mach-O shared libraries.

osSubsectionsViaSymbols :: OS -> Bool

platformUsesFrameworks :: Platform -> Bool