PrelNames

Documentation

data Unique

Unique identifier.

The type of unique identifiers that are used in many places in GHC for fast ordering and equality tests. You should generate these with the functions from the UniqSupply module

These are sometimes also referred to as "keys" in comments in GHC.

class Uniquable a where

Class of things that we can obtain a Unique from

Minimal complete definition

getUnique

Methods

getUnique :: a -> Unique

hasKey :: Uniquable a => a -> Unique -> Bool

allNameStrings :: [String]

itName :: Unique -> SrcSpan -> Name

mkUnboundName :: OccName -> Name

isUnboundName :: Name -> Bool

basicKnownKeyNames :: [Name]

genericTyConNames :: [Name]

pRELUDE :: Module

gHC_PRIM :: Module

gHC_TYPES :: Module

gHC_GENERICS :: Module

gHC_MAGIC :: Module

gHC_CLASSES :: Module

gHC_BASE :: Module

gHC_ENUM :: Module

gHC_GHCI :: Module

gHC_CSTRING :: Module

gHC_SHOW :: Module

gHC_READ :: Module

gHC_NUM :: Module

gHC_INTEGER_TYPE :: Module

gHC_NATURAL :: Module

gHC_LIST :: Module

gHC_TUPLE :: Module

dATA_TUPLE :: Module

dATA_EITHER :: Module

dATA_STRING :: Module

dATA_FOLDABLE :: Module

dATA_TRAVERSABLE :: Module

dATA_MONOID :: Module

dATA_SEMIGROUP :: Module

gHC_CONC :: Module

gHC_IO :: Module

gHC_IO_Exception :: Module

gHC_ST :: Module

gHC_ARR :: Module

gHC_STABLE :: Module

gHC_PTR :: Module

gHC_ERR :: Module

gHC_REAL :: Module

gHC_FLOAT :: Module

gHC_TOP_HANDLER :: Module

sYSTEM_IO :: Module

dYNAMIC :: Module

tYPEABLE :: Module

tYPEABLE_INTERNAL :: Module

gENERICS :: Module

rEAD_PREC :: Module

lEX :: Module

gHC_INT :: Module

gHC_WORD :: Module

mONAD :: Module

mONAD_FIX :: Module

mONAD_ZIP :: Module

mONAD_FAIL :: Module

aRROW :: Module

cONTROL_APPLICATIVE :: Module

gHC_DESUGAR :: Module

rANDOM :: Module

gHC_EXTS :: Module

cONTROL_EXCEPTION_BASE :: Module

gHC_TYPELITS :: Module

gHC_TYPENATS :: Module

dATA_TYPE_EQUALITY :: Module

dATA_COERCE :: Module

gHC_PARR' :: Module

gHC_SRCLOC :: Module

gHC_STACK :: Module

gHC_STACK_TYPES :: Module

gHC_STATICPTR :: Module

gHC_STATICPTR_INTERNAL :: Module

gHC_FINGERPRINT_TYPE :: Module

gHC_OVER_LABELS :: Module

gHC_RECORDS :: Module

mAIN :: Module

rOOT_MAIN :: Module

mkInteractiveModule :: Int -> Module

pRELUDE_NAME :: ModuleName

mAIN_NAME :: ModuleName

dATA_ARRAY_PARALLEL_NAME :: ModuleName

dATA_ARRAY_PARALLEL_PRIM_NAME :: ModuleName

mkPrimModule :: FastString -> Module

mkIntegerModule :: FastString -> Module

mkBaseModule :: FastString -> Module

mkBaseModule_ :: ModuleName -> Module

mkThisGhcModule :: FastString -> Module

mkThisGhcModule_ :: ModuleName -> Module

mkMainModule :: FastString -> Module

mkMainModule_ :: ModuleName -> Module

main_RDR_Unqual :: RdrName

forall_tv_RDR :: RdrName

dot_tv_RDR :: RdrName

eq_RDR :: RdrName

ge_RDR :: RdrName

ne_RDR :: RdrName

le_RDR :: RdrName

lt_RDR :: RdrName

gt_RDR :: RdrName

compare_RDR :: RdrName

ltTag_RDR :: RdrName

eqTag_RDR :: RdrName

gtTag_RDR :: RdrName

eqClass_RDR :: RdrName

numClass_RDR :: RdrName

ordClass_RDR :: RdrName

enumClass_RDR :: RdrName

monadClass_RDR :: RdrName

map_RDR :: RdrName

append_RDR :: RdrName

foldr_RDR :: RdrName

build_RDR :: RdrName

returnM_RDR :: RdrName

bindM_RDR :: RdrName

failM_RDR_preMFP :: RdrName

failM_RDR :: RdrName

left_RDR :: RdrName

right_RDR :: RdrName

fromEnum_RDR :: RdrName

toEnum_RDR :: RdrName

enumFrom_RDR :: RdrName

enumFromTo_RDR :: RdrName

enumFromThen_RDR :: RdrName

enumFromThenTo_RDR :: RdrName

ratioDataCon_RDR :: RdrName

plusInteger_RDR :: RdrName

timesInteger_RDR :: RdrName

ioDataCon_RDR :: RdrName

eqString_RDR :: RdrName

unpackCString_RDR :: RdrName

unpackCStringFoldr_RDR :: RdrName

unpackCStringUtf8_RDR :: RdrName

newStablePtr_RDR :: RdrName

bindIO_RDR :: RdrName

returnIO_RDR :: RdrName

fromInteger_RDR :: RdrName

fromRational_RDR :: RdrName

minus_RDR :: RdrName

times_RDR :: RdrName

plus_RDR :: RdrName

toInteger_RDR :: RdrName

toRational_RDR :: RdrName

fromIntegral_RDR :: RdrName

stringTy_RDR :: RdrName

fromString_RDR :: RdrName

fromList_RDR :: RdrName

fromListN_RDR :: RdrName

toList_RDR :: RdrName

compose_RDR :: RdrName

not_RDR :: RdrName

getTag_RDR :: RdrName

succ_RDR :: RdrName

pred_RDR :: RdrName

minBound_RDR :: RdrName

maxBound_RDR :: RdrName

and_RDR :: RdrName

range_RDR :: RdrName

inRange_RDR :: RdrName

index_RDR :: RdrName

unsafeIndex_RDR :: RdrName

unsafeRangeSize_RDR :: RdrName

readList_RDR :: RdrName

readListDefault_RDR :: RdrName

readListPrec_RDR :: RdrName

readListPrecDefault_RDR :: RdrName

readPrec_RDR :: RdrName

parens_RDR :: RdrName

choose_RDR :: RdrName

lexP_RDR :: RdrName

expectP_RDR :: RdrName

punc_RDR :: RdrName

ident_RDR :: RdrName

symbol_RDR :: RdrName

step_RDR :: RdrName

alt_RDR :: RdrName

reset_RDR :: RdrName

prec_RDR :: RdrName

pfail_RDR :: RdrName

showList_RDR :: RdrName

showList___RDR :: RdrName

showsPrec_RDR :: RdrName

shows_RDR :: RdrName

showString_RDR :: RdrName

showSpace_RDR :: RdrName

showCommaSpace_RDR :: RdrName

showParen_RDR :: RdrName

undefined_RDR :: RdrName

error_RDR :: RdrName

u1DataCon_RDR :: RdrName

par1DataCon_RDR :: RdrName

rec1DataCon_RDR :: RdrName

k1DataCon_RDR :: RdrName

m1DataCon_RDR :: RdrName

l1DataCon_RDR :: RdrName

r1DataCon_RDR :: RdrName

prodDataCon_RDR :: RdrName

comp1DataCon_RDR :: RdrName

unPar1_RDR :: RdrName

unRec1_RDR :: RdrName

unK1_RDR :: RdrName

unComp1_RDR :: RdrName

from_RDR :: RdrName

from1_RDR :: RdrName

to_RDR :: RdrName

to1_RDR :: RdrName

datatypeName_RDR :: RdrName

moduleName_RDR :: RdrName

packageName_RDR :: RdrName

isNewtypeName_RDR :: RdrName

conName_RDR :: RdrName

conFixity_RDR :: RdrName

conIsRecord_RDR :: RdrName

selName_RDR :: RdrName

prefixDataCon_RDR :: RdrName

infixDataCon_RDR :: RdrName

leftAssocDataCon_RDR :: RdrName

rightAssocDataCon_RDR :: RdrName

notAssocDataCon_RDR :: RdrName

uAddrDataCon_RDR :: RdrName

uCharDataCon_RDR :: RdrName

uDoubleDataCon_RDR :: RdrName

uFloatDataCon_RDR :: RdrName

uIntDataCon_RDR :: RdrName

uWordDataCon_RDR :: RdrName

uAddrHash_RDR :: RdrName

uCharHash_RDR :: RdrName

uDoubleHash_RDR :: RdrName

uFloatHash_RDR :: RdrName

uIntHash_RDR :: RdrName

uWordHash_RDR :: RdrName

fmap_RDR :: RdrName

replace_RDR :: RdrName

pure_RDR :: RdrName

ap_RDR :: RdrName

liftA2_RDR :: RdrName

foldable_foldr_RDR :: RdrName

foldMap_RDR :: RdrName

traverse_RDR :: RdrName

mempty_RDR :: RdrName

mappend_RDR :: RdrName

eqTyCon_RDR :: RdrName

varQual_RDR :: Module -> FastString -> RdrName

tcQual_RDR :: Module -> FastString -> RdrName

clsQual_RDR :: Module -> FastString -> RdrName

dataQual_RDR :: Module -> FastString -> RdrName

wildCardName :: Name

runMainIOName :: Name

orderingTyConName :: Name

ltDataConName :: Name

eqDataConName :: Name

gtDataConName :: Name

specTyConName :: Name

eitherTyConName :: Name

leftDataConName :: Name

rightDataConName :: Name

v1TyConName :: Name

u1TyConName :: Name

par1TyConName :: Name

rec1TyConName :: Name

k1TyConName :: Name

m1TyConName :: Name

sumTyConName :: Name

prodTyConName :: Name

compTyConName :: Name

rTyConName :: Name

dTyConName :: Name

cTyConName :: Name

sTyConName :: Name

rec0TyConName :: Name

d1TyConName :: Name

c1TyConName :: Name

s1TyConName :: Name

noSelTyConName :: Name

repTyConName :: Name

rep1TyConName :: Name

uRecTyConName :: Name

uAddrTyConName :: Name

uCharTyConName :: Name

uDoubleTyConName :: Name

uFloatTyConName :: Name

uIntTyConName :: Name

uWordTyConName :: Name

prefixIDataConName :: Name

infixIDataConName :: Name

leftAssociativeDataConName :: Name

rightAssociativeDataConName :: Name

notAssociativeDataConName :: Name

sourceUnpackDataConName :: Name

sourceNoUnpackDataConName :: Name

noSourceUnpackednessDataConName :: Name

sourceLazyDataConName :: Name

sourceStrictDataConName :: Name

noSourceStrictnessDataConName :: Name

decidedLazyDataConName :: Name

decidedStrictDataConName :: Name

decidedUnpackDataConName :: Name

metaDataDataConName :: Name

metaConsDataConName :: Name

metaSelDataConName :: Name

divIntName :: Name

modIntName :: Name

unpackCStringName :: Name

unpackCStringFoldrName :: Name

unpackCStringUtf8Name :: Name

eqStringName :: Name

inlineIdName :: Name

fmapName :: Name

eqClassName :: Name

eqName :: Name

ordClassName :: Name

geName :: Name

functorClassName :: Name

monadClassName :: Name

thenMName :: Name

bindMName :: Name

returnMName :: Name

failMName_preMFP :: Name

monadFailClassName :: Name

failMName :: Name

applicativeClassName :: Name

pureAName :: Name

apAName :: Name

thenAName :: Name

foldableClassName :: Name

traversableClassName :: Name

semigroupClassName :: Name

sappendName :: Name

monoidClassName :: Name

memptyName :: Name

mappendName :: Name

mconcatName :: Name

joinMName :: Name

alternativeClassName :: Name

joinMIdKey :: Unique

apAClassOpKey :: Unique

pureAClassOpKey :: Unique

thenAClassOpKey :: Unique

alternativeClassKey :: Unique

groupWithName :: Name

fromStringName :: Name

otherwiseIdName :: Name

foldrName :: Name

buildName :: Name

augmentName :: Name

mapName :: Name

appendName :: Name

assertName :: Name

breakpointName :: Name

breakpointCondName :: Name

breakpointAutoName :: Name

opaqueTyConName :: Name

breakpointJumpName :: Name

breakpointCondJumpName :: Name

breakpointAutoJumpName :: Name

fstName :: Name

sndName :: Name

numClassName :: Name

fromIntegerName :: Name

minusName :: Name

negateName :: Name

integerTyConName :: Name

mkIntegerName :: Name

integerSDataConName :: Name

integerToWord64Name :: Name

integerToInt64Name :: Name

word64ToIntegerName :: Name

int64ToIntegerName :: Name

plusIntegerName :: Name

timesIntegerName :: Name

smallIntegerName :: Name

wordToIntegerName :: Name

integerToWordName :: Name

integerToIntName :: Name

minusIntegerName :: Name

negateIntegerName :: Name

eqIntegerPrimName :: Name

neqIntegerPrimName :: Name

absIntegerName :: Name

signumIntegerName :: Name

leIntegerPrimName :: Name

gtIntegerPrimName :: Name

ltIntegerPrimName :: Name

geIntegerPrimName :: Name

compareIntegerName :: Name

quotRemIntegerName :: Name

divModIntegerName :: Name

quotIntegerName :: Name

remIntegerName :: Name

divIntegerName :: Name

modIntegerName :: Name

floatFromIntegerName :: Name

doubleFromIntegerName :: Name

encodeFloatIntegerName :: Name

encodeDoubleIntegerName :: Name

decodeDoubleIntegerName :: Name

gcdIntegerName :: Name

lcmIntegerName :: Name

andIntegerName :: Name

orIntegerName :: Name

xorIntegerName :: Name

complementIntegerName :: Name

shiftLIntegerName :: Name

shiftRIntegerName :: Name

bitIntegerName :: Name

naturalTyConName :: Name

naturalFromIntegerName :: Name

rationalTyConName :: Name

ratioTyConName :: Name

ratioDataConName :: Name

realClassName :: Name

integralClassName :: Name

realFracClassName :: Name

fractionalClassName :: Name

fromRationalName :: Name

toIntegerName :: Name

toRationalName :: Name

fromIntegralName :: Name

realToFracName :: Name

floatingClassName :: Name

realFloatClassName :: Name

rationalToFloatName :: Name

rationalToDoubleName :: Name

ixClassName :: Name

trModuleTyConName :: Name

trModuleDataConName :: Name

trNameTyConName :: Name

trNameSDataConName :: Name

trNameDDataConName :: Name

trTyConTyConName :: Name

trTyConDataConName :: Name

kindRepTyConName :: Name

kindRepTyConAppDataConName :: Name

kindRepVarDataConName :: Name

kindRepAppDataConName :: Name

kindRepFunDataConName :: Name

kindRepTYPEDataConName :: Name

kindRepTypeLitSDataConName :: Name

kindRepTypeLitDDataConName :: Name

typeLitSortTyConName :: Name

typeLitSymbolDataConName :: Name

typeLitNatDataConName :: Name

typeableClassName :: Name

typeRepTyConName :: Name

someTypeRepTyConName :: Name

someTypeRepDataConName :: Name

mkTrConName :: Name

mkTrAppName :: Name

mkTrFunName :: Name

typeRepIdName :: Name

typeNatTypeRepName :: Name

typeSymbolTypeRepName :: Name

trGhcPrimModuleName :: Name

starKindRepName :: Name

starArrStarKindRepName :: Name

starArrStarArrStarKindRepName :: Name

errorMessageTypeErrorFamName :: Name

typeErrorTextDataConName :: Name

typeErrorAppendDataConName :: Name

typeErrorVAppendDataConName :: Name

typeErrorShowTypeDataConName :: Name

toDynName :: Name

dataClassName :: Name

assertErrorName :: Name

enumClassName :: Name

enumFromName :: Name

enumFromToName :: Name

enumFromThenName :: Name

enumFromThenToName :: Name

boundedClassName :: Name

concatName :: Name

filterName :: Name

zipName :: Name

isListClassName :: Name

fromListName :: Name

fromListNName :: Name

toListName :: Name

showClassName :: Name

readClassName :: Name

genClassName :: Name

gen1ClassName :: Name

datatypeClassName :: Name

constructorClassName :: Name

selectorClassName :: Name

genericClassNames :: [Name]

ghciIoClassName :: Name

ghciStepIoMName :: Name

ioTyConName :: Name

ioDataConName :: Name

thenIOName :: Name

bindIOName :: Name

returnIOName :: Name

failIOName :: Name

printName :: Name

int8TyConName :: Name

int16TyConName :: Name

int32TyConName :: Name

int64TyConName :: Name

word16TyConName :: Name

word32TyConName :: Name

word64TyConName :: Name

ptrTyConName :: Name

funPtrTyConName :: Name

stablePtrTyConName :: Name

newStablePtrName :: Name

monadFixClassName :: Name

mfixName :: Name

arrAName :: Name

composeAName :: Name

firstAName :: Name

appAName :: Name

choiceAName :: Name

loopAName :: Name

guardMName :: Name

liftMName :: Name

mzipName :: Name

toAnnotationWrapperName :: Name

monadPlusClassName :: Name

randomClassName :: Name

randomGenClassName :: Name

isStringClassName :: Name

knownNatClassName :: Name

knownSymbolClassName :: Name

isLabelClassName :: Name

ipClassName :: Name

hasFieldClassName :: Name

callStackTyConName :: Name

emptyCallStackName :: Name

pushCallStackName :: Name

srcLocDataConName :: Name

pLUGINS :: Module

pluginTyConName :: Name

frontendPluginTyConName :: Name

makeStaticName :: Name

staticPtrInfoTyConName :: Name

staticPtrInfoDataConName :: Name

staticPtrTyConName :: Name

staticPtrDataConName :: Name

fromStaticPtrName :: Name

fingerprintDataConName :: Name

eqTyConName :: Name

varQual :: Module -> FastString -> Unique -> Name

tcQual :: Module -> FastString -> Unique -> Name

clsQual :: Module -> FastString -> Unique -> Name

dcQual :: Module -> FastString -> Unique -> Name

mk_known_key_name :: NameSpace -> Module -> FastString -> Unique -> Name

boundedClassKey :: Unique

enumClassKey :: Unique

eqClassKey :: Unique

floatingClassKey :: Unique

fractionalClassKey :: Unique

integralClassKey :: Unique

monadClassKey :: Unique

dataClassKey :: Unique

functorClassKey :: Unique

numClassKey :: Unique

ordClassKey :: Unique

readClassKey :: Unique

realClassKey :: Unique

realFloatClassKey :: Unique

realFracClassKey :: Unique

showClassKey :: Unique

ixClassKey :: Unique

typeableClassKey :: Unique

typeable1ClassKey :: Unique

typeable2ClassKey :: Unique

typeable3ClassKey :: Unique

typeable4ClassKey :: Unique

typeable5ClassKey :: Unique

typeable6ClassKey :: Unique

typeable7ClassKey :: Unique

monadFixClassKey :: Unique

monadFailClassKey :: Unique

monadPlusClassKey :: Unique

randomClassKey :: Unique

randomGenClassKey :: Unique

isStringClassKey :: Unique

applicativeClassKey :: Unique

foldableClassKey :: Unique

traversableClassKey :: Unique

genClassKey :: Unique

gen1ClassKey :: Unique

datatypeClassKey :: Unique

constructorClassKey :: Unique

selectorClassKey :: Unique

knownNatClassNameKey :: Unique

knownSymbolClassNameKey :: Unique

ghciIoClassKey :: Unique

isLabelClassNameKey :: Unique

semigroupClassKey :: Unique

monoidClassKey :: Unique

ipClassKey :: Unique

hasFieldClassNameKey :: Unique

addrPrimTyConKey :: Unique

arrayPrimTyConKey :: Unique

arrayArrayPrimTyConKey :: Unique

boolTyConKey :: Unique

byteArrayPrimTyConKey :: Unique

charPrimTyConKey :: Unique

charTyConKey :: Unique

doublePrimTyConKey :: Unique

doubleTyConKey :: Unique

floatPrimTyConKey :: Unique

floatTyConKey :: Unique

funTyConKey :: Unique

intPrimTyConKey :: Unique

intTyConKey :: Unique

int8TyConKey :: Unique

int16TyConKey :: Unique

int32PrimTyConKey :: Unique

int32TyConKey :: Unique

int64PrimTyConKey :: Unique

int64TyConKey :: Unique

integerTyConKey :: Unique

naturalTyConKey :: Unique

listTyConKey :: Unique

foreignObjPrimTyConKey :: Unique

maybeTyConKey :: Unique

weakPrimTyConKey :: Unique

mutableArrayPrimTyConKey :: Unique

mutableArrayArrayPrimTyConKey :: Unique

mutableByteArrayPrimTyConKey :: Unique

orderingTyConKey :: Unique

mVarPrimTyConKey :: Unique

ratioTyConKey :: Unique

rationalTyConKey :: Unique

realWorldTyConKey :: Unique

stablePtrPrimTyConKey :: Unique

stablePtrTyConKey :: Unique

eqTyConKey :: Unique

heqTyConKey :: Unique

smallArrayPrimTyConKey :: Unique

smallMutableArrayPrimTyConKey :: Unique

statePrimTyConKey :: Unique

stableNamePrimTyConKey :: Unique

stableNameTyConKey :: Unique

mutVarPrimTyConKey :: Unique

ioTyConKey :: Unique

wordPrimTyConKey :: Unique

wordTyConKey :: Unique

word8TyConKey :: Unique

word16TyConKey :: Unique

word32PrimTyConKey :: Unique

word32TyConKey :: Unique

word64PrimTyConKey :: Unique

word64TyConKey :: Unique

liftedConKey :: Unique

unliftedConKey :: Unique

anyBoxConKey :: Unique

kindConKey :: Unique

boxityConKey :: Unique

typeConKey :: Unique

threadIdPrimTyConKey :: Unique

bcoPrimTyConKey :: Unique

ptrTyConKey :: Unique

funPtrTyConKey :: Unique

tVarPrimTyConKey :: Unique

eqPrimTyConKey :: Unique

eqReprPrimTyConKey :: Unique

eqPhantPrimTyConKey :: Unique

voidPrimTyConKey :: Unique

compactPrimTyConKey :: Unique

parrTyConKey :: Unique

objectTyConKey :: Unique

eitherTyConKey :: Unique

liftedTypeKindTyConKey :: Unique

tYPETyConKey :: Unique

constraintKindTyConKey :: Unique

starKindTyConKey :: Unique

unicodeStarKindTyConKey :: Unique

runtimeRepTyConKey :: Unique

vecCountTyConKey :: Unique

vecElemTyConKey :: Unique

pluginTyConKey :: Unique

frontendPluginTyConKey :: Unique

unknownTyConKey :: Unique

unknown1TyConKey :: Unique

unknown2TyConKey :: Unique

unknown3TyConKey :: Unique

opaqueTyConKey :: Unique

v1TyConKey :: Unique

u1TyConKey :: Unique

par1TyConKey :: Unique

rec1TyConKey :: Unique

k1TyConKey :: Unique

m1TyConKey :: Unique

sumTyConKey :: Unique

prodTyConKey :: Unique

compTyConKey :: Unique

rTyConKey :: Unique

dTyConKey :: Unique

cTyConKey :: Unique

sTyConKey :: Unique

rec0TyConKey :: Unique

d1TyConKey :: Unique

c1TyConKey :: Unique

s1TyConKey :: Unique

noSelTyConKey :: Unique

repTyConKey :: Unique

rep1TyConKey :: Unique

uRecTyConKey :: Unique

uAddrTyConKey :: Unique

uCharTyConKey :: Unique

uDoubleTyConKey :: Unique

uFloatTyConKey :: Unique

uIntTyConKey :: Unique

uWordTyConKey :: Unique

typeNatKindConNameKey :: Unique

typeSymbolKindConNameKey :: Unique

typeNatAddTyFamNameKey :: Unique

typeNatMulTyFamNameKey :: Unique

typeNatExpTyFamNameKey :: Unique

typeNatLeqTyFamNameKey :: Unique

typeNatSubTyFamNameKey :: Unique

typeSymbolCmpTyFamNameKey :: Unique

typeNatCmpTyFamNameKey :: Unique

errorMessageTypeErrorFamKey :: Unique

ntTyConKey :: Unique

coercibleTyConKey :: Unique

proxyPrimTyConKey :: Unique

specTyConKey :: Unique

anyTyConKey :: Unique

staticPtrTyConKey :: Unique

staticPtrInfoTyConKey :: Unique

callStackTyConKey :: Unique

typeRepTyConKey :: Unique

someTypeRepTyConKey :: Unique

someTypeRepDataConKey :: Unique

typeSymbolAppendFamNameKey :: Unique

charDataConKey :: Unique

consDataConKey :: Unique

doubleDataConKey :: Unique

falseDataConKey :: Unique

floatDataConKey :: Unique

intDataConKey :: Unique

integerSDataConKey :: Unique

nilDataConKey :: Unique

ratioDataConKey :: Unique

stableNameDataConKey :: Unique

trueDataConKey :: Unique

wordDataConKey :: Unique

word8DataConKey :: Unique

ioDataConKey :: Unique

integerDataConKey :: Unique

heqDataConKey :: Unique

coercibleDataConKey :: Unique

nothingDataConKey :: Unique

justDataConKey :: Unique

crossDataConKey :: Unique

inlDataConKey :: Unique

inrDataConKey :: Unique

genUnitDataConKey :: Unique

parrDataConKey :: Unique

leftDataConKey :: Unique

rightDataConKey :: Unique

ltDataConKey :: Unique

eqDataConKey :: Unique

gtDataConKey :: Unique

staticPtrDataConKey :: Unique

staticPtrInfoDataConKey :: Unique

fingerprintDataConKey :: Unique

srcLocDataConKey :: Unique

trTyConTyConKey :: Unique

trTyConDataConKey :: Unique

trModuleTyConKey :: Unique

trModuleDataConKey :: Unique

trNameTyConKey :: Unique

trNameSDataConKey :: Unique

trNameDDataConKey :: Unique

trGhcPrimModuleKey :: Unique

kindRepTyConKey :: Unique

typeLitSortTyConKey :: Unique

typeErrorTextDataConKey :: Unique

typeErrorAppendDataConKey :: Unique

typeErrorVAppendDataConKey :: Unique

typeErrorShowTypeDataConKey :: Unique

prefixIDataConKey :: Unique

infixIDataConKey :: Unique

leftAssociativeDataConKey :: Unique

rightAssociativeDataConKey :: Unique

notAssociativeDataConKey :: Unique

sourceUnpackDataConKey :: Unique

sourceNoUnpackDataConKey :: Unique

noSourceUnpackednessDataConKey :: Unique

sourceLazyDataConKey :: Unique

sourceStrictDataConKey :: Unique

noSourceStrictnessDataConKey :: Unique

decidedLazyDataConKey :: Unique

decidedStrictDataConKey :: Unique

decidedUnpackDataConKey :: Unique

metaDataDataConKey :: Unique

metaConsDataConKey :: Unique

metaSelDataConKey :: Unique

vecRepDataConKey :: Unique

tupleRepDataConKey :: Unique

sumRepDataConKey :: Unique

runtimeRepSimpleDataConKeys :: [Unique]

liftedRepDataConKey :: Unique

vecCountDataConKeys :: [Unique]

vecElemDataConKeys :: [Unique]

kindRepTyConAppDataConKey :: Unique

kindRepVarDataConKey :: Unique

kindRepAppDataConKey :: Unique

kindRepFunDataConKey :: Unique

kindRepTYPEDataConKey :: Unique

kindRepTypeLitSDataConKey :: Unique

kindRepTypeLitDDataConKey :: Unique

typeLitSymbolDataConKey :: Unique

typeLitNatDataConKey :: Unique

wildCardKey :: Unique

absentErrorIdKey :: Unique

augmentIdKey :: Unique

appendIdKey :: Unique

buildIdKey :: Unique

errorIdKey :: Unique

foldrIdKey :: Unique

recSelErrorIdKey :: Unique

seqIdKey :: Unique

irrefutPatErrorIdKey :: Unique

eqStringIdKey :: Unique

noMethodBindingErrorIdKey :: Unique

nonExhaustiveGuardsErrorIdKey :: Unique

runtimeErrorIdKey :: Unique

patErrorIdKey :: Unique

voidPrimIdKey :: Unique

realWorldPrimIdKey :: Unique

recConErrorIdKey :: Unique

unpackCStringUtf8IdKey :: Unique

unpackCStringAppendIdKey :: Unique

unpackCStringFoldrIdKey :: Unique

unpackCStringIdKey :: Unique

typeErrorIdKey :: Unique

divIntIdKey :: Unique

modIntIdKey :: Unique

unsafeCoerceIdKey :: Unique

concatIdKey :: Unique

filterIdKey :: Unique

zipIdKey :: Unique

bindIOIdKey :: Unique

returnIOIdKey :: Unique

newStablePtrIdKey :: Unique

printIdKey :: Unique

failIOIdKey :: Unique

nullAddrIdKey :: Unique

voidArgIdKey :: Unique

fstIdKey :: Unique

sndIdKey :: Unique

otherwiseIdKey :: Unique

assertIdKey :: Unique

mkIntegerIdKey :: Unique

smallIntegerIdKey :: Unique

wordToIntegerIdKey :: Unique

integerToWordIdKey :: Unique

integerToIntIdKey :: Unique

integerToWord64IdKey :: Unique

integerToInt64IdKey :: Unique

word64ToIntegerIdKey :: Unique

int64ToIntegerIdKey :: Unique

plusIntegerIdKey :: Unique

timesIntegerIdKey :: Unique

minusIntegerIdKey :: Unique

negateIntegerIdKey :: Unique

eqIntegerPrimIdKey :: Unique

neqIntegerPrimIdKey :: Unique

absIntegerIdKey :: Unique

signumIntegerIdKey :: Unique

leIntegerPrimIdKey :: Unique

gtIntegerPrimIdKey :: Unique

ltIntegerPrimIdKey :: Unique

geIntegerPrimIdKey :: Unique

compareIntegerIdKey :: Unique

quotRemIntegerIdKey :: Unique

divModIntegerIdKey :: Unique

quotIntegerIdKey :: Unique

remIntegerIdKey :: Unique

divIntegerIdKey :: Unique

modIntegerIdKey :: Unique

floatFromIntegerIdKey :: Unique

doubleFromIntegerIdKey :: Unique

encodeFloatIntegerIdKey :: Unique

encodeDoubleIntegerIdKey :: Unique

decodeDoubleIntegerIdKey :: Unique

gcdIntegerIdKey :: Unique

lcmIntegerIdKey :: Unique

andIntegerIdKey :: Unique

orIntegerIdKey :: Unique

xorIntegerIdKey :: Unique

complementIntegerIdKey :: Unique

shiftLIntegerIdKey :: Unique

shiftRIntegerIdKey :: Unique

rootMainKey :: Unique

runMainKey :: Unique

thenIOIdKey :: Unique

lazyIdKey :: Unique

assertErrorIdKey :: Unique

oneShotKey :: Unique

runRWKey :: Unique

breakpointIdKey :: Unique

breakpointCondIdKey :: Unique

breakpointAutoIdKey :: Unique

breakpointJumpIdKey :: Unique

breakpointCondJumpIdKey :: Unique

breakpointAutoJumpIdKey :: Unique

inlineIdKey :: Unique

noinlineIdKey :: Unique

mapIdKey :: Unique

groupWithIdKey :: Unique

dollarIdKey :: Unique

coercionTokenIdKey :: Unique

rationalToFloatIdKey :: Unique

rationalToDoubleIdKey :: Unique

unmarshalObjectIdKey :: Unique

marshalObjectIdKey :: Unique

marshalStringIdKey :: Unique

unmarshalStringIdKey :: Unique

checkDotnetResNameIdKey :: Unique

undefinedKey :: Unique

magicDictKey :: Unique

coerceKey :: Unique

unboundKey :: Unique

fromIntegerClassOpKey :: Unique

minusClassOpKey :: Unique

fromRationalClassOpKey :: Unique

enumFromClassOpKey :: Unique

enumFromThenClassOpKey :: Unique

enumFromToClassOpKey :: Unique

enumFromThenToClassOpKey :: Unique

eqClassOpKey :: Unique

geClassOpKey :: Unique

negateClassOpKey :: Unique

failMClassOpKey_preMFP :: Unique

bindMClassOpKey :: Unique

thenMClassOpKey :: Unique

returnMClassOpKey :: Unique

fmapClassOpKey :: Unique

mfixIdKey :: Unique

failMClassOpKey :: Unique

arrAIdKey :: Unique

composeAIdKey :: Unique

firstAIdKey :: Unique

appAIdKey :: Unique

choiceAIdKey :: Unique

loopAIdKey :: Unique

fromStringClassOpKey :: Unique

toAnnotationWrapperIdKey :: Unique

fromIntegralIdKey :: Unique

realToFracIdKey :: Unique

toIntegerClassOpKey :: Unique

toRationalClassOpKey :: Unique

guardMIdKey :: Unique

liftMIdKey :: Unique

mzipIdKey :: Unique

ghciStepIoMClassOpKey :: Unique

isListClassKey :: Unique

fromListClassOpKey :: Unique

fromListNClassOpKey :: Unique

toListClassOpKey :: Unique

proxyHashKey :: Unique

mkTyConKey :: Unique

mkTrConKey :: Unique

mkTrAppKey :: Unique

mkTrFunKey :: Unique

typeNatTypeRepKey :: Unique

typeSymbolTypeRepKey :: Unique

typeRepIdKey :: Unique

trTYPEKey :: Unique

trTYPE'PtrRepLiftedKey :: Unique

trRuntimeRepKey :: Unique

tr'PtrRepLiftedKey :: Unique

starKindRepKey :: Unique

starArrStarKindRepKey :: Unique

starArrStarArrStarKindRepKey :: Unique

toDynIdKey :: Unique

bitIntegerIdKey :: Unique

heqSCSelIdKey :: Unique

coercibleSCSelIdKey :: Unique

sappendClassOpKey :: Unique

memptyClassOpKey :: Unique

mappendClassOpKey :: Unique

mconcatClassOpKey :: Unique

emptyCallStackKey :: Unique

pushCallStackKey :: Unique

fromStaticPtrClassOpKey :: Unique

makeStaticKey :: Unique

naturalFromIntegerIdKey :: Unique

numericClassKeys :: [Unique]

fractionalClassKeys :: [Unique]

standardClassKeys :: [Unique]

derivableClassKeys :: [Unique]

interactiveClassNames :: [Name]

interactiveClassKeys :: [Unique]

pretendNameIsInScope :: Name -> Bool

int8X16PrimTyConKey :: Unique

int16X8PrimTyConKey :: Unique

int32X4PrimTyConKey :: Unique

int64X2PrimTyConKey :: Unique

int8X32PrimTyConKey :: Unique

int16X16PrimTyConKey :: Unique

int32X8PrimTyConKey :: Unique

int64X4PrimTyConKey :: Unique

int8X64PrimTyConKey :: Unique

int16X32PrimTyConKey :: Unique

int32X16PrimTyConKey :: Unique

int64X8PrimTyConKey :: Unique

word8X16PrimTyConKey :: Unique

word16X8PrimTyConKey :: Unique

word32X4PrimTyConKey :: Unique

word64X2PrimTyConKey :: Unique

word8X32PrimTyConKey :: Unique

word16X16PrimTyConKey :: Unique

word32X8PrimTyConKey :: Unique

word64X4PrimTyConKey :: Unique

word8X64PrimTyConKey :: Unique

word16X32PrimTyConKey :: Unique

word32X16PrimTyConKey :: Unique

word64X8PrimTyConKey :: Unique

floatX4PrimTyConKey :: Unique

doubleX2PrimTyConKey :: Unique

floatX8PrimTyConKey :: Unique

doubleX4PrimTyConKey :: Unique

floatX16PrimTyConKey :: Unique

doubleX8PrimTyConKey :: Unique