RegAlloc.Graph.Spill

Description

When there aren't enough registers to hold all the vregs we have to spill some of those vregs to slots on the stack. This module is used modify the code to use those slots.

Documentation

regSpill

Arguments

:: Instruction instr

=> Platform

-> [LiveCmmDecl statics instr]
the code

-> UniqSet Int
available stack slots

-> UniqSet VirtualReg
the regs to spill

-> UniqSM ([LiveCmmDecl statics instr], UniqSet Int, SpillStats)

Spill all these virtual regs to stack slots.

TODO: See if we can split some of the live ranges instead of just globally spilling the virtual reg. This might make the spill cleaner's job easier.

TODO: On CISCy x86 and x86_64 we don't nessesarally have to add a mov instruction when making spills. If an instr is using a spilled virtual we may be able to address the spill slot directly.

data SpillStats

Spiller statistics. Tells us what registers were spilled.

Constructors

SpillStats

Fields

spillStoreLoad :: UniqFM (Reg, Int, Int)

accSpillSL :: (Reg, Int, Int) -> (Reg, Int, Int) -> (Reg, Int, Int)

Add a spill/reload count to a stats record for a register.