RegAlloc.Linear.SPARC.FreeRegs

Description

Free regs map for SPARC

Documentation

data FreeRegs

Constructors

FreeRegs !Word32 !Word32 !Word32

noFreeRegs :: FreeRegs

A reg map where no regs are free to be allocated.

initFreeRegs :: Platform -> FreeRegs

The initial set of free regs.

getFreeRegs :: RegClass -> FreeRegs -> [RealReg]

Get all the free registers of this class.

allocateReg :: Platform -> RealReg -> FreeRegs -> FreeRegs

Grab a register.

releaseReg :: Platform -> RealReg -> FreeRegs -> FreeRegs

Release a register from allocation. The register liveness information says that most regs die after a C call, but we still don't want to allocate to some of them.

bitMask :: Int -> Word32

showFreeRegs :: FreeRegs -> String