RnExpr

Documentation

rnLExpr :: LHsExpr GhcPs -> RnM (LHsExpr GhcRn, FreeVars)

rnExpr :: HsExpr GhcPs -> RnM (HsExpr GhcRn, FreeVars)

rnStmts

Arguments

:: Outputable (body GhcPs)

=> HsStmtContext Name

-> (Located (body GhcPs) -> RnM (Located (body GhcRn), FreeVars))
How to rename the body of each statement (e.g. rnLExpr)

-> [LStmt GhcPs (Located (body GhcPs))]
Statements

-> ([Name] -> RnM (thing, FreeVars))
if these statements scope over something, this renames it and returns the result.

-> RnM (([LStmt GhcRn (Located (body GhcRn))], thing), FreeVars)

Rename some Stmts