RnFixity

Documentation

type MiniFixityEnv = FastStringEnv (Located Fixity)

addLocalFixities :: MiniFixityEnv -> [Name] -> RnM a -> RnM a

lookupFixityRn :: Name -> RnM Fixity

lookupFixityRn_help :: Name -> RnM (Bool, Fixity)

lookupFixityRn_help returns (True, fixity) if it finds a Fixity in a local environment or from an interface file. Otherwise, it returns (False, fixity) (e.g., for unbound Names or Names without user-supplied fixity declarations).

lookupFieldFixityRn :: AmbiguousFieldOcc GhcRn -> RnM Fixity

Look up the fixity of a (possibly ambiguous) occurrence of a record field selector. We use lookupFixityRn' so that we can specifiy the OccName as the field label, which might be different to the OccName of the selector Name if DuplicateRecordFields is in use (Trac #1173). If there are multiple possible selectors with different fixities, generate an error.

lookupTyFixityRn :: Located Name -> RnM Fixity