SPARC.CodeGen

Documentation

cmmTopCodeGen :: RawCmmDecl -> NatM [NatCmmDecl CmmStatics Instr]

Top level code generation

generateJumpTableForInstr :: DynFlags -> Instr -> Maybe (NatCmmDecl CmmStatics Instr)

type InstrBlock = OrdList Instr

InstrBlocks are the insn sequences generated by the insn selectors. They are really trees of insns to facilitate fast appending, where a left-to-right traversal yields the insns in the correct order.