SPARC.Regs

Documentation

showReg :: RegNo -> String

Get the standard name for the register with this number.

virtualRegSqueeze :: RegClass -> VirtualReg -> Int

regSqueeze_class reg Calculuate the maximum number of register colors that could be denied to a node of this class due to having this reg as a neighbour.

realRegSqueeze :: RegClass -> RealReg -> Int

classOfRealReg :: RealReg -> RegClass

allRealRegs :: [RealReg]

All the allocatable registers in the machine, including register pairs.

gReg :: Int -> RegNo

Get the regno for this sort of reg

iReg :: Int -> RegNo

Get the regno for this sort of reg

lReg :: Int -> RegNo

Get the regno for this sort of reg

oReg :: Int -> RegNo

Get the regno for this sort of reg

fReg :: Int -> RegNo

Get the regno for this sort of reg

fp :: Reg

Some specific regs used by the code generator.

sp :: Reg

Some specific regs used by the code generator.

g0 :: Reg

Some specific regs used by the code generator.

g1 :: Reg

Some specific regs used by the code generator.

g2 :: Reg

Some specific regs used by the code generator.

o0 :: Reg

Some specific regs used by the code generator.

o1 :: Reg

Some specific regs used by the code generator.

f0 :: Reg

Some specific regs used by the code generator.

f1 :: Reg

Some specific regs used by the code generator.

f6 :: Reg

Some specific regs used by the code generator.

f8 :: Reg

Some specific regs used by the code generator.

f22 :: Reg

Some specific regs used by the code generator.

f26 :: Reg

Some specific regs used by the code generator.

f27 :: Reg

Some specific regs used by the code generator.

allocatableRegs :: [RealReg]

Produce the second-half-of-a-double register given the first half.

All the regs that the register allocator can allocate to, with the the fixed use regs removed.

argRegs :: RegNo -> [Reg]

The registers to place arguments for function calls, for some number of arguments.

allArgRegs :: [Reg]

All all the regs that could possibly be returned by argRegs

callClobberedRegs :: [Reg]

mkVirtualReg :: Unique -> Format -> VirtualReg

Make a virtual reg with this format.

regDotColor :: RealReg -> SDoc