StgCmmForeign

Documentation

cgForeignCall :: ForeignCall -> [StgArg] -> Type -> FCode ReturnKind

emit code for a foreign call, and return the results to the sequel.

emitPrimCall :: [CmmFormal] -> CallishMachOp -> [CmmActual] -> FCode ()

emitCCall :: [(CmmFormal, ForeignHint)] -> CmmExpr -> [(CmmActual, ForeignHint)] -> FCode ()

emitForeignCall :: Safety -> [CmmFormal] -> ForeignTarget -> [CmmActual] -> FCode ReturnKind

emitSaveThreadState :: FCode ()

saveThreadState :: MonadUnique m => DynFlags -> m CmmAGraph

Produce code to save the current thread state to CurrentTSO

emitLoadThreadState :: FCode ()

loadThreadState :: MonadUnique m => DynFlags -> m CmmAGraph

Produce code to load the current thread state from CurrentTSO

emitOpenNursery :: FCode ()

emitCloseNursery :: FCode ()