StgCmmLayout

Documentation

mkArgDescr :: DynFlags -> [Id] -> ArgDescr

emitCall :: (Convention, Convention) -> CmmExpr -> [CmmExpr] -> FCode ReturnKind

emitCall conv fun args makes a call to the entry-code of fun, using the call/return convention conv, passing args, and returning the results to the current sequel.

emitReturn :: [CmmExpr] -> FCode ReturnKind

Return multiple values to the sequel

If the sequel is Return

    return (x,y)

If the sequel is AssignTo [p,q]

   p=x; q=y;

adjustHpBackwards :: FCode ()

emitClosureProcAndInfoTable :: Bool -> Id -> LambdaFormInfo -> CmmInfoTable -> [NonVoid Id] -> ((Int, LocalReg, [LocalReg]) -> FCode ()) -> FCode ()

emitClosureAndInfoTable :: CmmInfoTable -> Convention -> [LocalReg] -> FCode () -> FCode ()

slowCall :: CmmExpr -> [StgArg] -> FCode ReturnKind

directCall :: Convention -> CLabel -> RepArity -> [StgArg] -> FCode ReturnKind

mkVirtHeapOffsets :: DynFlags -> Bool -> [NonVoid (PrimRep, a)] -> (WordOff, WordOff, [(NonVoid a, ByteOff)])

mkVirtConstrOffsets :: DynFlags -> [NonVoid (PrimRep, a)] -> (WordOff, WordOff, [(NonVoid a, ByteOff)])

Just like mkVirtHeapOffsets, but for constructors

mkVirtConstrSizes :: DynFlags -> [NonVoid PrimRep] -> (WordOff, WordOff)

Just like mkVirtConstrOffsets, but used when we don't have the actual arguments. Useful when e.g. generating info tables; we just need to know sizes of pointer and non-pointer fields.

getHpRelOffset :: VirtualHpOffset -> FCode CmmExpr

data ArgRep

Constructors

P

N

L

V

F

D

V16

V32

V64

toArgRep :: PrimRep -> ArgRep

argRepSizeW :: DynFlags -> ArgRep -> WordOff