StgCmmProf

Documentation

initCostCentres :: CollectedCCs -> FCode ()

ccType :: DynFlags -> CmmType

ccsType :: DynFlags -> CmmType

mkCCostCentre :: CostCentre -> CmmLit

mkCCostCentreStack :: CostCentreStack -> CmmLit

dynProfHdr :: DynFlags -> CmmExpr -> [CmmExpr]

Profiling header words in a dynamic closure

profDynAlloc :: SMRep -> CmmExpr -> FCode ()

Record the allocation of a closure. The CmmExpr is the cost centre stack to which to attribute the allocation.

profAlloc :: CmmExpr -> CmmExpr -> FCode ()

Record the allocation of a closure (size is given by a CmmExpr) The size must be in words, because the allocation counter in a CCS counts in words.

staticProfHdr :: DynFlags -> CostCentreStack -> [CmmLit]

The profiling header words in a static closure

initUpdFrameProf :: CmmExpr -> FCode ()

Initialise the profiling field of an update frame

enterCostCentreThunk :: CmmExpr -> FCode ()

enterCostCentreFun :: CostCentreStack -> CmmExpr -> FCode ()

costCentreFrom :: DynFlags -> CmmExpr -> CmmExpr

storeCurCCS :: CmmExpr -> CmmAGraph

emitSetCCC :: CostCentre -> Bool -> Bool -> FCode ()

saveCurrentCostCentre :: FCode (Maybe LocalReg)

restoreCurrentCostCentre :: Maybe LocalReg -> FCode ()

ldvEnter :: CmmExpr -> FCode ()

ldvEnterClosure :: ClosureInfo -> CmmReg -> FCode ()

Called when a closure is entered, marks the closure as having been "used". The closure is not an "inherently used" one. The closure is not IND because that is not considered for LDV profiling.

ldvRecordCreate :: CmmExpr -> FCode ()