StgSyn

Documentation

data StgArg

Constructors

StgVarArg Id

StgLitArg Literal

data GenStgTopBinding pass

A top-level binding.

Constructors

StgTopLifted (GenStgBinding pass)

StgTopStringLit Id ByteString

data GenStgBinding pass

Constructors

StgNonRec (BinderP pass) (GenStgRhs pass)

StgRec [(BinderP pass, GenStgRhs pass)]

data GenStgExpr pass

Constructors

StgApp Id [StgArg]

StgLit Literal

StgConApp DataCon [StgArg] [Type]

StgOpApp StgOp [StgArg] Type

StgLam (NonEmpty (BinderP pass)) StgExpr

StgCase (GenStgExpr pass) (BinderP pass) AltType [GenStgAlt pass]

StgLet (XLet pass) (GenStgBinding pass) (GenStgExpr pass)

StgLetNoEscape (XLetNoEscape pass) (GenStgBinding pass) (GenStgExpr pass)

StgTick (Tickish Id) (GenStgExpr pass)

data GenStgRhs pass

Constructors

StgRhsClosure

Fields

(XRhsClosure pass)

Extension point for non-global free var list just before CodeGen.

CostCentreStack

CCS to be attached (default is CurrentCCS)

!UpdateFlag

ReEntrant | Updatable | SingleEntry

[BinderP pass]

arguments; if empty, then not a function; as above, order is important.

(GenStgExpr pass)

body

StgRhsCon CostCentreStack DataCon [StgArg]

type GenStgAlt pass = (AltCon, [BinderP pass], GenStgExpr pass)

data AltType

Constructors

PolyAlt

MultiValAlt Int

AlgAlt TyCon

PrimAlt PrimRep

data StgPass

Used as a data type index for the stgSyn AST

Constructors

Vanilla

LiftLams

CodeGen

type family BinderP (pass :: StgPass)

type family XRhsClosure (pass :: StgPass)

type family XLet (pass :: StgPass)

type family XLetNoEscape (pass :: StgPass)

data NoExtSilent

Like NoExt, but with an Outputable instance that returns empty.

noExtSilent :: NoExtSilent

Used when constructing a term with an unused extension point that should not appear in pretty-printed output at all.

type OutputablePass pass = (Outputable (XLet pass), Outputable (XLetNoEscape pass), Outputable (XRhsClosure pass), OutputableBndr (BinderP pass))

data UpdateFlag

Constructors

ReEntrant

Updatable

SingleEntry

isUpdatable :: UpdateFlag -> Bool

type StgTopBinding = GenStgTopBinding 'Vanilla

type StgBinding = GenStgBinding 'Vanilla

type StgExpr = GenStgExpr 'Vanilla

type StgRhs = GenStgRhs 'Vanilla

type StgAlt = GenStgAlt 'Vanilla

type CgStgTopBinding = GenStgTopBinding 'CodeGen

type CgStgBinding = GenStgBinding 'CodeGen

type CgStgExpr = GenStgExpr 'CodeGen

type CgStgRhs = GenStgRhs 'CodeGen

type CgStgAlt = GenStgAlt 'CodeGen

type LlStgTopBinding = GenStgTopBinding 'LiftLams

type LlStgBinding = GenStgBinding 'LiftLams

type LlStgExpr = GenStgExpr 'LiftLams

type LlStgRhs = GenStgRhs 'LiftLams

type LlStgAlt = GenStgAlt 'LiftLams

type InStgArg = StgArg

type InStgTopBinding = StgTopBinding

type InStgBinding = StgBinding

type InStgExpr = StgExpr

type InStgRhs = StgRhs

type InStgAlt = StgAlt

type OutStgArg = StgArg

type OutStgTopBinding = StgTopBinding

type OutStgBinding = StgBinding

type OutStgExpr = StgExpr

type OutStgRhs = StgRhs

type OutStgAlt = StgAlt

data StgOp

Constructors

StgPrimOp PrimOp

StgPrimCallOp PrimCall

StgFCallOp ForeignCall Unique

topStgBindHasCafRefs :: GenStgTopBinding pass -> Bool

stgArgHasCafRefs :: StgArg -> Bool

stgRhsArity :: StgRhs -> Int

isDllConApp :: DynFlags -> Module -> DataCon -> [StgArg] -> Bool

Does this constructor application refer to anything in a different *Windows* DLL? If so, we can't allocate it statically

stgArgType :: StgArg -> Type

Type of an StgArg

Very half baked because we have lost the type arguments.

stripStgTicksTop :: (Tickish Id -> Bool) -> GenStgExpr p -> ([Tickish Id], GenStgExpr p)

Strip ticks of a given type from an STG expression

stgCaseBndrInScope

Arguments

:: AltType

-> Bool
unarised?

-> Bool

Given an alt type and whether the program is unarised, return whether the case binder is in scope.

Case binders of unboxed tuple or unboxed sum type always dead after the unariser has run. See Note [Post-unarisation invariants].

pprStgBinding :: StgBinding -> SDoc

pprGenStgTopBindings :: OutputablePass pass => [GenStgTopBinding pass] -> SDoc

pprStgTopBindings :: [StgTopBinding] -> SDoc