TcEvidence

Documentation

data HsWrapper

Constructors

WpHole

WpCompose HsWrapper HsWrapper

WpFun HsWrapper HsWrapper TcType SDoc

WpCast TcCoercionR

WpEvLam EvVar

WpEvApp EvTerm

WpTyLam TyVar

WpTyApp KindOrType

WpLet TcEvBinds

(<.>) :: HsWrapper -> HsWrapper -> HsWrapper

mkWpTyApps :: [Type] -> HsWrapper

mkWpEvApps :: [EvTerm] -> HsWrapper

mkWpEvVarApps :: [EvVar] -> HsWrapper

mkWpTyLams :: [TyVar] -> HsWrapper

mkWpLams :: [Var] -> HsWrapper

mkWpLet :: TcEvBinds -> HsWrapper

mkWpCastN :: TcCoercionN -> HsWrapper

mkWpCastR :: TcCoercionR -> HsWrapper

collectHsWrapBinders :: HsWrapper -> ([Var], HsWrapper)

mkWpFun :: HsWrapper -> HsWrapper -> TcType -> TcType -> SDoc -> HsWrapper

mkWpFuns :: [(TcType, HsWrapper)] -> TcType -> HsWrapper -> SDoc -> HsWrapper

mkWpFuns [(ty1, wrap1), (ty2, wrap2)] ty_res wrap_res, where wrap1 :: ty1 "->" ty1' and wrap2 :: ty2 "->" ty2', wrap3 :: ty3 "->" ty3' and ty_res is either ty3 or ty3', gives a wrapper (ty1' -> ty2' -> ty3) "->" (ty1 -> ty2 -> ty3'). Notice that the result wrapper goes the other way round to all the others. This is a result of sub-typing contravariance. The SDoc is a description of what you were doing when you called mkWpFuns.

idHsWrapper :: HsWrapper

isIdHsWrapper :: HsWrapper -> Bool

pprHsWrapper :: HsWrapper -> (Bool -> SDoc) -> SDoc

data TcEvBinds

Constructors

TcEvBinds EvBindsVar

EvBinds (Bag EvBind)

data EvBindsVar

Constructors

EvBindsVar

Fields

ebv_uniq :: Unique

ebv_binds :: IORef EvBindMap

ebv_tcvs :: IORef CoVarSet

NoEvBindsVar

Fields

ebv_uniq :: Unique

ebv_tcvs :: IORef CoVarSet

newtype EvBindMap

Constructors

EvBindMap

Fields

ev_bind_varenv :: DVarEnv EvBind

emptyEvBindMap :: EvBindMap

extendEvBinds :: EvBindMap -> EvBind -> EvBindMap

lookupEvBind :: EvBindMap -> EvVar -> Maybe EvBind

evBindMapBinds :: EvBindMap -> Bag EvBind

foldEvBindMap :: (EvBind -> a -> a) -> a -> EvBindMap -> a

filterEvBindMap :: (EvBind -> Bool) -> EvBindMap -> EvBindMap

isEmptyEvBindMap :: EvBindMap -> Bool

data EvBind

Constructors

EvBind

Fields

eb_lhs :: EvVar

eb_rhs :: EvTerm

eb_is_given :: Bool

emptyTcEvBinds :: TcEvBinds

isEmptyTcEvBinds :: TcEvBinds -> Bool

mkGivenEvBind :: EvVar -> EvTerm -> EvBind

mkWantedEvBind :: EvVar -> EvTerm -> EvBind

evBindVar :: EvBind -> EvVar

isNoEvBindsVar :: EvBindsVar -> Bool

data EvTerm

Constructors

EvExpr EvExpr

EvTypeable Type EvTypeable

EvFun

Fields

et_tvs :: [TyVar]

et_given :: [EvVar]

et_binds :: TcEvBinds

et_body :: EvVar

type EvExpr = CoreExpr

evId :: EvId -> EvExpr

Any sort of evidence Id, including coercions

evCoercion :: TcCoercion -> EvTerm

evCast :: EvExpr -> TcCoercion -> EvTerm

d |> co

evDFunApp :: DFunId -> [Type] -> [EvExpr] -> EvTerm

evSelector :: Id -> [Type] -> [EvExpr] -> EvExpr

mkEvCast :: EvExpr -> TcCoercion -> EvTerm

evVarsOfTerm :: EvTerm -> VarSet

mkEvScSelectors :: Class -> [TcType] -> [(TcPredType, EvExpr)]

evTypeable :: Type -> EvTypeable -> EvTerm

findNeededEvVars :: EvBindMap -> VarSet -> VarSet

evTermCoercion :: EvTerm -> TcCoercion

evTermCoercion_maybe :: EvTerm -> Maybe TcCoercion

data EvCallStack

Evidence for CallStack implicit parameters.

Constructors

EvCsEmpty

EvCsPushCall Name RealSrcSpan EvExpr
EvCsPushCall name loc stk represents a call to name, occurring at loc, in a calling context stk.

data EvTypeable

Instructions on how to make a Typeable dictionary. See Note [Typeable evidence terms]

Constructors

EvTypeableTyCon TyCon [EvTerm]
Dictionary for Typeable T where T is a type constructor with all of its kind variables saturated. The [EvTerm] is Typeable evidence for the applied kinds..

EvTypeableTyApp EvTerm EvTerm
Dictionary for Typeable (s t), given a dictionaries for s and t.

EvTypeableTrFun EvTerm EvTerm
Dictionary for Typeable (s -> t), given a dictionaries for s and t.

EvTypeableTyLit EvTerm
Dictionary for a type literal, e.g. Typeable "foo" or Typeable 3 The EvTerm is evidence of, e.g., KnownNat 3 (see Trac #10348)

type TcCoercion = Coercion

type TcCoercionR = CoercionR

type TcCoercionN = CoercionN

type TcCoercionP = CoercionP

data CoercionHole

A coercion to be filled in by the type-checker. See Note [Coercion holes]

data Role

Constructors

Nominal

Representational

Phantom

data LeftOrRight

Constructors

CLeft

CRight

pickLR :: LeftOrRight -> (a, a) -> a

mkTcReflCo :: Role -> TcType -> TcCoercion

mkTcNomReflCo :: TcType -> TcCoercionN

mkTcRepReflCo :: TcType -> TcCoercionR

mkTcTyConAppCo :: Role -> TyCon -> [TcCoercion] -> TcCoercion

mkTcAppCo :: TcCoercion -> TcCoercionN -> TcCoercion

mkTcFunCo :: Role -> TcCoercion -> TcCoercion -> TcCoercion

mkTcAxInstCo :: Role -> CoAxiom br -> BranchIndex -> [TcType] -> [TcCoercion] -> TcCoercion

mkTcUnbranchedAxInstCo :: CoAxiom Unbranched -> [TcType] -> [TcCoercion] -> TcCoercionR

mkTcForAllCo :: TyVar -> TcCoercionN -> TcCoercion -> TcCoercion

mkTcForAllCos :: [(TyVar, TcCoercionN)] -> TcCoercion -> TcCoercion

mkTcSymCo :: TcCoercion -> TcCoercion

mkTcTransCo :: TcCoercion -> TcCoercion -> TcCoercion

mkTcNthCo :: Role -> Int -> TcCoercion -> TcCoercion

mkTcLRCo :: LeftOrRight -> TcCoercion -> TcCoercion

mkTcSubCo :: TcCoercionN -> TcCoercionR

maybeTcSubCo :: EqRel -> TcCoercion -> TcCoercion

tcDowngradeRole :: Role -> Role -> TcCoercion -> TcCoercion

mkTcAxiomRuleCo :: CoAxiomRule -> [TcCoercion] -> TcCoercionR

mkTcCoherenceLeftCo :: TcCoercion -> TcCoercionN -> TcCoercion

mkTcCoherenceRightCo :: TcCoercion -> TcCoercionN -> TcCoercion

mkTcPhantomCo :: TcCoercionN -> TcType -> TcType -> TcCoercionP

mkTcKindCo :: TcCoercion -> TcCoercionN

tcCoercionKind :: TcCoercion -> Pair TcType

coVarsOfTcCo :: TcCoercion -> TcTyCoVarSet

mkTcCoVarCo :: CoVar -> TcCoercion

isTcReflCo :: TcCoercion -> Bool

isTcReflexiveCo :: TcCoercion -> Bool

This version does a slow check, calculating the related types and seeing if they are equal.

tcCoercionRole :: TcCoercion -> Role

unwrapIP :: Type -> CoercionR

Create a Expr that unwraps an implicit-parameter or overloaded-label dictionary to expose the underlying value. We expect the Expr to have the form `IP sym ty` or `IsLabel sym ty`, and return a Expr `co :: IP sym ty ~ ty` or `co :: IsLabel sym ty ~ Proxy# sym -> ty`. See also Note [Type-checking overloaded labels] in TcExpr.

wrapIP :: Type -> CoercionR

Create a Expr that wraps a value in an implicit-parameter dictionary. See unwrapIP.