TcRnTypes

Documentation

type TcRnIf a b = IOEnv (Env a b)

type TcRn = TcRnIf TcGblEnv TcLclEnv

type TcM = TcRn

Historical "type-checking monad" (now it's just TcRn).

type RnM = TcRn

Historical "renaming monad" (now it's just TcRn).

type IfM lcl = TcRnIf IfGblEnv lcl

type IfL = IfM IfLclEnv

type IfG = IfM ()

type TcRef a = IORef a

Type alias for IORef; the convention is we'll use this for mutable bits of data in TcGblEnv which are updated during typechecking and returned at the end.

data Env gbl lcl

Constructors

Env

Fields

env_top :: !HscEnv

env_um :: !Char

env_gbl :: gbl

env_lcl :: lcl

data TcGblEnv

TcGblEnv describes the top-level of the module at the point at which the typechecker is finished work. It is this structure that is handed on to the desugarer For state that needs to be updated during the typechecking phase and returned at end, use a TcRef (= IORef).

Constructors

TcGblEnv

Fields

tcg_mod :: Module

Module being compiled

tcg_semantic_mod :: Module

If a signature, the backing module See also Note [Identity versus semantic module]

tcg_src :: HscSource

What kind of module (regular Haskell, hs-boot, hsig)

tcg_rdr_env :: GlobalRdrEnv

Top level envt; used during renaming

tcg_default :: Maybe [Type]

Types used for defaulting. Nothing => no default decl

tcg_fix_env :: FixityEnv

Just for things in this module

tcg_field_env :: RecFieldEnv

Just for things in this module See Note [The interactive package] in HscTypes

tcg_type_env :: TypeEnv

Global type env for the module we are compiling now. All TyCons and Classes (for this module) end up in here right away, along with their derived constructors, selectors.

(Ids defined in this module start in the local envt, though they move to the global envt during zonking)

NB: for what "things in this module" means, see Note [The interactive package] in HscTypes

tcg_type_env_var :: TcRef TypeEnv

tcg_inst_env :: !InstEnv

Instance envt for all home-package modules; Includes the dfuns in tcg_insts NB. BangPattern is to fix a leak, see #15111

tcg_fam_inst_env :: !FamInstEnv

Ditto for family instances NB. BangPattern is to fix a leak, see #15111

tcg_ann_env :: AnnEnv

And for annotations

tcg_exports :: [AvailInfo]

What is exported

tcg_imports :: ImportAvails

Information about what was imported from where, including things bound in this module. Also store Safe Haskell info here about transitive trusted package requirements.

There are not many uses of this field, so you can grep for all them.

The ImportAvails records information about the following things:

All of the modules you directly imported (tcRnImports)

The orphans (only!) of all imported modules in a GHCi session (runTcInteractive)

The module that instantiated a signature

Each of the signatures that merged in

It is used in the following ways: - imp_orphs is used to determine what orphan modules should be visible in the context (tcVisibleOrphanMods) - imp_finsts is used to determine what family instances should be visible (tcExtendLocalFamInstEnv) - To resolve the meaning of the export list of a module (tcRnExports) - imp_mods is used to compute usage info (mkIfaceTc, deSugar) - imp_trust_own_pkg is used for Safe Haskell in interfaces (mkIfaceTc, as well as in HscMain) - To create the Dependencies field in interface (mkDependencies)

tcg_dus :: DefUses

tcg_used_gres :: TcRef [GlobalRdrElt]

tcg_keep :: TcRef NameSet

tcg_th_used :: TcRef Bool

True = Template Haskell syntax used.

We need this so that we can generate a dependency on the Template Haskell package, because the desugarer is going to emit loads of references to TH symbols. The reference is implicit rather than explicit, so we have to zap a mutable variable.

tcg_th_splice_used :: TcRef Bool

True = A Template Haskell splice was used.

Splices disable recompilation avoidance (see #481)

tcg_th_top_level_locs :: TcRef (Set RealSrcSpan)

Locations of the top-level splices; used for providing details on scope in error messages for out-of-scope variables

tcg_dfun_n :: TcRef OccSet

Allows us to choose unique DFun names.

tcg_merged :: [(Module, Fingerprint)]

The requirements we merged with; we always have to recompile if any of these changed.

tcg_rn_exports :: Maybe [(Located (IE GhcRn), Avails)]

tcg_rn_imports :: [LImportDecl GhcRn]

tcg_rn_decls :: Maybe (HsGroup GhcRn)

Renamed decls, maybe. Nothing = Don't retain renamed decls.

tcg_dependent_files :: TcRef [FilePath]

dependencies from addDependentFile

tcg_th_topdecls :: TcRef [LHsDecl GhcPs]

Top-level declarations from addTopDecls

tcg_th_foreign_files :: TcRef [(ForeignSrcLang, FilePath)]

Foreign files emitted from TH.

tcg_th_topnames :: TcRef NameSet

Exact names bound in top-level declarations in tcg_th_topdecls

tcg_th_modfinalizers :: TcRef [(TcLclEnv, ThModFinalizers)]

Template Haskell module finalizers.

They can use particular local environments.

tcg_th_coreplugins :: TcRef [String]

Core plugins added by Template Haskell code.

tcg_th_state :: TcRef (Map TypeRep Dynamic)

tcg_th_remote_state :: TcRef (Maybe (ForeignRef (IORef QState)))

Template Haskell state

tcg_ev_binds :: Bag EvBind

tcg_tr_module :: Maybe Id

tcg_binds :: LHsBinds GhcTc

tcg_sigs :: NameSet

tcg_imp_specs :: [LTcSpecPrag]

tcg_warns :: Warnings

tcg_anns :: [Annotation]

tcg_tcs :: [TyCon]

tcg_insts :: [ClsInst]

tcg_fam_insts :: [FamInst]

tcg_rules :: [LRuleDecl GhcTc]

tcg_fords :: [LForeignDecl GhcTc]

tcg_patsyns :: [PatSyn]

tcg_doc_hdr :: Maybe LHsDocString

Maybe Haddock header docs

tcg_hpc :: !AnyHpcUsage

True if any part of the prog uses hpc instrumentation. NB. BangPattern is to fix a leak, see #15111

tcg_self_boot :: SelfBootInfo

Whether this module has a corresponding hi-boot file

tcg_main :: Maybe Name

The Name of the main function, if this module is the main module.

tcg_safeInfer :: TcRef (Bool, WarningMessages)

Has the typechecker inferred this module as -XSafe (Safe Haskell) See Note [Safe Haskell Overlapping Instances Implementation], although this is used for more than just that failure case.

tcg_tc_plugins :: [TcPluginSolver]

A list of user-defined plugins for the constraint solver.

tcg_hf_plugins :: [HoleFitPlugin]

A list of user-defined plugins for hole fit suggestions.

tcg_top_loc :: RealSrcSpan

The RealSrcSpan this module came from

tcg_static_wc :: TcRef WantedConstraints

Wanted constraints of static forms. See Note [Constraints in static forms].

tcg_complete_matches :: [CompleteMatch]

Tracking indices for cost centre annotations

tcg_cc_st :: TcRef CostCentreState

data TcLclEnv

Constructors

TcLclEnv

Fields

tcl_loc :: RealSrcSpan

tcl_ctxt :: [ErrCtxt]

tcl_tclvl :: TcLevel

tcl_th_ctxt :: ThStage

tcl_th_bndrs :: ThBindEnv

tcl_arrow_ctxt :: ArrowCtxt

tcl_rdr :: LocalRdrEnv

tcl_env :: TcTypeEnv

tcl_bndrs :: TcBinderStack

tcl_lie :: TcRef WantedConstraints

tcl_errs :: TcRef Messages

setLclEnvTcLevel :: TcLclEnv -> TcLevel -> TcLclEnv

getLclEnvTcLevel :: TcLclEnv -> TcLevel

setLclEnvLoc :: TcLclEnv -> RealSrcSpan -> TcLclEnv

getLclEnvLoc :: TcLclEnv -> RealSrcSpan

data IfGblEnv

Constructors

IfGblEnv

Fields

if_doc :: SDoc

if_rec_types :: Maybe (Module, IfG TypeEnv)

data IfLclEnv

Constructors

IfLclEnv

Fields

if_mod :: Module

if_boot :: Bool

if_loc :: SDoc

if_nsubst :: Maybe NameShape

if_implicits_env :: Maybe TypeEnv

if_tv_env :: FastStringEnv TyVar

if_id_env :: FastStringEnv Id

tcVisibleOrphanMods :: TcGblEnv -> ModuleSet

data FrontendResult

FrontendResult describes the result of running the frontend of a Haskell module. Usually, you'll get a FrontendTypecheck, since running the frontend involves typechecking a program, but for an hs-boot merge you'll just get a ModIface, since no actual typechecking occurred.

This data type really should be in HscTypes, but it needs to have a TcGblEnv which is only defined here.

Constructors

FrontendTypecheck TcGblEnv

type ErrCtxt = (Bool, TidyEnv -> TcM (TidyEnv, MsgDoc))

type RecFieldEnv = NameEnv [FieldLabel]

pushErrCtxt :: CtOrigin -> ErrCtxt -> CtLoc -> CtLoc

pushErrCtxtSameOrigin :: ErrCtxt -> CtLoc -> CtLoc

data ImportAvails

ImportAvails summarises what was imported from where, irrespective of whether the imported things are actually used or not. It is used:

when processing the export list,

when constructing usage info for the interface file,

to identify the list of directly imported modules for initialisation purposes and for optimised overlap checking of family instances,

when figuring out what things are really unused

Constructors

ImportAvails

Fields

imp_mods :: ImportedMods

Domain is all directly-imported modules

See the documentation on ImportedModsVal in HscTypes for the meaning of the fields.

We need a full ModuleEnv rather than a ModuleNameEnv here, because we might be importing modules of the same name from different packages. (currently not the case, but might be in the future).

imp_dep_mods :: ModuleNameEnv (ModuleName, IsBootInterface)

Home-package modules needed by the module being compiled

It doesn't matter whether any of these dependencies are actually used when compiling the module; they are listed if they are below it at all. For example, suppose M imports A which imports X. Then compiling M might not need to consult X.hi, but X is still listed in M's dependencies.

imp_dep_pkgs :: Set InstalledUnitId

Packages needed by the module being compiled, whether directly, or via other modules in this package, or via modules imported from other packages.

imp_trust_pkgs :: Set InstalledUnitId

This is strictly a subset of imp_dep_pkgs and records the packages the current module needs to trust for Safe Haskell compilation to succeed. A package is required to be trusted if we are dependent on a trustworthy module in that package. While perhaps making imp_dep_pkgs a tuple of (UnitId, Bool) where True for the bool indicates the package is required to be trusted is the more logical design, doing so complicates a lot of code not concerned with Safe Haskell. See Note [RnNames . Tracking Trust Transitively]

imp_trust_own_pkg :: Bool

Do we require that our own package is trusted? This is to handle efficiently the case where a Safe module imports a Trustworthy module that resides in the same package as it. See Note [RnNames . Trust Own Package]

imp_orphs :: [Module]

Orphan modules below us in the import tree (and maybe including us for imported modules)

imp_finsts :: [Module]

Family instance modules below us in the import tree (and maybe including us for imported modules)

emptyImportAvails :: ImportAvails

plusImportAvails :: ImportAvails -> ImportAvails -> ImportAvails

Union two ImportAvails

This function is a key part of Import handling, basically for each import we create a separate ImportAvails structure and then union them all together with this function.

data WhereFrom

Constructors

ImportByUser IsBootInterface

ImportBySystem

ImportByPlugin

mkModDeps :: [(ModuleName, IsBootInterface)] -> ModuleNameEnv (ModuleName, IsBootInterface)

modDepsElts :: ModuleNameEnv (ModuleName, IsBootInterface) -> [(ModuleName, IsBootInterface)]

type TcTypeEnv = NameEnv TcTyThing

type TcBinderStack = [TcBinder]

data TcBinder

Constructors

TcIdBndr TcId TopLevelFlag

TcIdBndr_ExpType Name ExpType TopLevelFlag

TcTvBndr Name TyVar

data TcTyThing

A typecheckable thing available in a local context. Could be AGlobal TyThing, but also lexically scoped variables, etc. See TcEnv for how to retrieve a TyThing given a Name.

Constructors

AGlobal TyThing

ATcId

Fields

tct_id :: TcId

tct_info :: IdBindingInfo

ATyVar Name TcTyVar

ATcTyCon TyCon

APromotionErr PromotionErr

data PromotionErr

Constructors

TyConPE

ClassPE

FamDataConPE

ConstrainedDataConPE PredType

PatSynPE

RecDataConPE

NoDataKindsTC

NoDataKindsDC

data IdBindingInfo

IdBindingInfo describes how an Id is bound.

It is used for the following purposes: a) for static forms in TcExpr.checkClosedInStaticForm and b) to figure out when a nested binding can be generalised, in TcBinds.decideGeneralisationPlan.

Constructors

NotLetBound

ClosedLet

NonClosedLet RhsNames ClosedTypeId

type ClosedTypeId = Bool

type RhsNames = NameSet

data IsGroupClosed

IsGroupClosed describes a group of mutually-recursive bindings

Constructors

IsGroupClosed (NameEnv RhsNames) ClosedTypeId

data SelfBootInfo

Constructors

NoSelfBoot

SelfBoot

Fields

sb_mds :: ModDetails

sb_tcs :: NameSet

pprTcTyThingCategory :: TcTyThing -> SDoc

pprPECategory :: PromotionErr -> SDoc

data CompleteMatch

A list of conlikes which represents a complete pattern match. These arise from COMPLETE signatures.

Constructors

CompleteMatch

Fields

completeMatchConLikes :: [Name]

The ConLikes that form a covering family (e.g. Nothing, Just)

completeMatchTyCon :: Name

The TyCon that they cover (e.g. Maybe)

type DsM = TcRnIf DsGblEnv DsLclEnv

data DsLclEnv

Constructors

DsLclEnv

Fields

dsl_meta :: DsMetaEnv

dsl_loc :: RealSrcSpan

dsl_delta :: Delta

data DsGblEnv

Constructors

DsGblEnv

Fields

ds_mod :: Module

ds_fam_inst_env :: FamInstEnv

ds_unqual :: PrintUnqualified

ds_msgs :: IORef Messages

ds_if_env :: (IfGblEnv, IfLclEnv)

ds_complete_matches :: CompleteMatchMap

ds_cc_st :: IORef CostCentreState

type DsMetaEnv = NameEnv DsMetaVal

data DsMetaVal

Constructors

DsBound Id

DsSplice (HsExpr GhcTc)

type CompleteMatchMap = UniqFM [CompleteMatch]

A map keyed by the completeMatchTyCon.

mkCompleteMatchMap :: [CompleteMatch] -> CompleteMatchMap

extendCompleteMatchMap :: CompleteMatchMap -> [CompleteMatch] -> CompleteMatchMap

data ThStage

Constructors

Splice SpliceType

RunSplice (TcRef [ForeignRef (Q ())])

Comp

Brack ThStage PendingStuff

data SpliceType

Constructors

Typed

Untyped

data PendingStuff

Constructors

RnPendingUntyped (TcRef [PendingRnSplice])

RnPendingTyped

TcPending (TcRef [PendingTcSplice]) (TcRef WantedConstraints)

topStage :: ThStage

topAnnStage :: ThStage

topSpliceStage :: ThStage

type ThLevel = Int

impLevel :: ThLevel

outerLevel :: ThLevel

thLevel :: ThStage -> ThLevel

data ForeignSrcLang

Foreign formats supported by GHC via TH

Constructors

LangC
C

LangCxx
C++

LangObjc
Objective C

LangObjcxx
Objective C++

LangAsm
Assembly language (.s)

RawObject
Object (.o)

data ArrowCtxt

Constructors

NoArrowCtxt

ArrowCtxt LocalRdrEnv (TcRef WantedConstraints)

type TcSigFun = Name -> Maybe TcSigInfo

data TcSigInfo

Constructors

TcIdSig TcIdSigInfo

TcPatSynSig TcPatSynInfo

data TcIdSigInfo

Constructors

CompleteSig

Fields

sig_bndr :: TcId

sig_ctxt :: UserTypeCtxt

sig_loc :: SrcSpan

PartialSig

Fields

psig_name :: Name

psig_hs_ty :: LHsSigWcType GhcRn

sig_ctxt :: UserTypeCtxt

sig_loc :: SrcSpan

data TcIdSigInst

Constructors

TISI

Fields

sig_inst_sig :: TcIdSigInfo

sig_inst_skols :: [(Name, TcTyVar)]

sig_inst_theta :: TcThetaType

sig_inst_tau :: TcSigmaType

sig_inst_wcs :: [(Name, TcTyVar)]

sig_inst_wcx :: Maybe TcType

data TcPatSynInfo

Constructors

TPSI

Fields

patsig_name :: Name

patsig_implicit_bndrs :: [TyVarBinder]

patsig_univ_bndrs :: [TyVar]

patsig_req :: TcThetaType

patsig_ex_bndrs :: [TyVar]

patsig_prov :: TcThetaType

patsig_body_ty :: TcSigmaType

isPartialSig :: TcIdSigInst -> Bool

hasCompleteSig :: TcSigFun -> Name -> Bool

No signature or a partial signature

type TcId = Id

type TcIdSet = IdSet

data Hole

An expression or type hole

Constructors

ExprHole UnboundVar
Either an out-of-scope variable or a "true" hole in an expression (TypedHoles)

TypeHole OccName
A hole in a type (PartialTypeSignatures)

holeOcc :: Hole -> OccName

data NameShape

A NameShape is a substitution on Names that can be used to refine the identities of a hole while we are renaming interfaces (see RnModIface). Specifically, a NameShape for ns_module_name A, defines a mapping from {A.T} (for some OccName T) to some arbitrary other Name.

The most intruiging thing about a NameShape, however, is how it's constructed. A NameShape is *implied* by the exported AvailInfos of the implementor of an interface: if an implementor of signature H exports M.T, you implicitly define a substitution from {H.T} to M.T. So a NameShape is computed from the list of AvailInfos that are exported by the implementation of a module, or successively merged together by the export lists of signatures which are joining together.

It's not the most obvious way to go about doing this, but it does seem to work!

NB: Can't boot this and put it in NameShape because then we start pulling in too many DynFlags things.

Constructors

NameShape

Fields

ns_mod_name :: ModuleName

ns_exports :: [AvailInfo]

ns_map :: OccEnv Name

removeBindingShadowing :: HasOccName a => [a] -> [a]

data TcPlugin

Constructors

forall s. TcPlugin

Fields

tcPluginInit :: TcPluginM s

Initialize plugin, when entering type-checker.

tcPluginSolve :: s -> TcPluginSolver

Solve some constraints. TODO: WRITE MORE DETAILS ON HOW THIS WORKS.

tcPluginStop :: s -> TcPluginM ()

Clean up after the plugin, when exiting the type-checker.

data TcPluginResult

Constructors

TcPluginContradiction [Ct]
The plugin found a contradiction. The returned constraints are removed from the inert set, and recorded as insoluble.

TcPluginOk [(EvTerm, Ct)] [Ct]
The first field is for constraints that were solved. These are removed from the inert set, and the evidence for them is recorded. The second field contains new work, that should be processed by the constraint solver.

type TcPluginSolver = [Ct] -> [Ct] -> [Ct] -> TcPluginM TcPluginResult

data TcPluginM a

runTcPluginM :: TcPluginM a -> EvBindsVar -> TcM a

unsafeTcPluginTcM :: TcM a -> TcPluginM a

This function provides an escape for direct access to the TcM monad. It should not be used lightly, and the provided TcPluginM API should be favoured instead.

getEvBindsTcPluginM :: TcPluginM EvBindsVar

Access the EvBindsVar carried by the TcPluginM during constraint solving. Returns Nothing if invoked during tcPluginInit or tcPluginStop.

type RoleAnnotEnv = NameEnv (LRoleAnnotDecl GhcRn)

emptyRoleAnnotEnv :: RoleAnnotEnv

mkRoleAnnotEnv :: [LRoleAnnotDecl GhcRn] -> RoleAnnotEnv

lookupRoleAnnot :: RoleAnnotEnv -> Name -> Maybe (LRoleAnnotDecl GhcRn)

getRoleAnnots :: [Name] -> RoleAnnotEnv -> [LRoleAnnotDecl GhcRn]