TyCoFVs

Documentation

tyCoVarsOfType :: Type -> TyCoVarSet

tyCoVarsOfTypeDSet :: Type -> DTyCoVarSet

tyCoFVsOfType that returns free variables of a type in a deterministic set. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in FV.

tyCoVarsOfTypes :: [Type] -> TyCoVarSet

tyCoVarsOfTypesDSet :: [Type] -> DTyCoVarSet

Returns free variables of types, including kind variables as a deterministic set. For type synonyms it does not expand the synonym.

exactTyCoVarsOfType :: Type -> TyCoVarSet

exactTyCoVarsOfTypes :: [Type] -> TyVarSet

tyCoFVsBndr :: TyCoVarBinder -> FV -> FV

tyCoFVsVarBndr :: Var -> FV -> FV

tyCoFVsVarBndrs :: [Var] -> FV -> FV

tyCoFVsOfType :: Type -> FV

The worker for tyCoFVsOfType and tyCoFVsOfTypeList. The previous implementation used unionVarSet which is O(n+m) and can make the function quadratic. It's exported, so that it can be composed with other functions that compute free variables. See Note [FV naming conventions] in FV.

Eta-expanded because that makes it run faster (apparently) See Note [FV eta expansion] in FV for explanation.

tyCoVarsOfTypeList :: Type -> [TyCoVar]

tyCoFVsOfType that returns free variables of a type in deterministic order. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in FV.

tyCoFVsOfTypes :: [Type] -> FV

tyCoVarsOfTypesList :: [Type] -> [TyCoVar]

Returns free variables of types, including kind variables as a deterministically ordered list. For type synonyms it does not expand the synonym.

tyCoVarsOfTypesSet :: TyVarEnv Type -> TyCoVarSet

Returns free variables of types, including kind variables as a non-deterministic set. For type synonyms it does not expand the synonym.

tyCoVarsOfCosSet :: CoVarEnv Coercion -> TyCoVarSet

coVarsOfType :: Type -> CoVarSet

coVarsOfTypes :: [Type] -> TyCoVarSet

coVarsOfCo :: Coercion -> CoVarSet

coVarsOfCos :: [Coercion] -> CoVarSet

tyCoVarsOfCo :: Coercion -> TyCoVarSet

tyCoVarsOfCos :: [Coercion] -> TyCoVarSet

tyCoVarsOfCoDSet :: Coercion -> DTyCoVarSet

Get a deterministic set of the vars free in a coercion

tyCoFVsOfCo :: Coercion -> FV

tyCoFVsOfCos :: [Coercion] -> FV

tyCoVarsOfCoList :: Coercion -> [TyCoVar]

tyCoVarsOfProv :: UnivCoProvenance -> TyCoVarSet

almostDevoidCoVarOfCo :: CoVar -> Coercion -> Bool

Given a covar and a coercion, returns True if covar is almost devoid in the coercion. That is, covar can only appear in Refl and GRefl. See last wrinkle in Note [Unused coercion variable in ForAllCo] in Coercion

injectiveVarsOfType

Arguments

:: Bool
Should we look under injective type families? See Note [Coverage condition for injective type families] in FamInst.

-> Type

-> FV

Returns the free variables of a Type that are in injective positions. Specifically, it finds the free variables while:

Expanding type synonyms

Ignoring the coercion in (ty |> co)

Ignoring the non-injective fields of a TyConApp

For example, if F is a non-injective type family, then:

injectiveTyVarsOf( Either c (Maybe (a, F b c)) ) = {a,c}

If injectiveVarsOfType ty = itvs, then knowing ty fixes itvs. More formally, if a is in injectiveVarsOfType ty and S1(ty) ~ S2(ty), then S1(a) ~ S2(a), where S1 and S2 are arbitrary substitutions.

See Note [When does a tycon application need an explicit kind signature?].

injectiveVarsOfTypes

Arguments

:: Bool
look under injective type families? See Note [Coverage condition for injective type families] in FamInst.

-> [Type]

-> FV

Returns the free variables of a Type that are in injective positions. Specifically, it finds the free variables while:

Expanding type synonyms

Ignoring the coercion in (ty |> co)

Ignoring the non-injective fields of a TyConApp

See Note [When does a tycon application need an explicit kind signature?].

invisibleVarsOfType :: Type -> FV

Returns the set of variables that are used invisibly anywhere within the given type. A variable will be included even if it is used both visibly and invisibly. An invisible use site includes: * In the kind of a variable * In the kind of a bound variable in a forall * In a coercion * In a Specified or Inferred argument to a function See Note [VarBndrs, TyCoVarBinders, TyConBinders, and visibility] in TyCoRep

invisibleVarsOfTypes :: [Type] -> FV

Like invisibleVarsOfType, but for many types.

noFreeVarsOfType :: Type -> Bool

Returns True if this type has no free variables. Should be the same as isEmptyVarSet . tyCoVarsOfType, but faster in the non-forall case.

noFreeVarsOfTypes :: [Type] -> Bool

noFreeVarsOfCo :: Coercion -> Bool

Returns True if this coercion has no free variables. Should be the same as isEmptyVarSet . tyCoVarsOfCo, but faster in the non-forall case.

mkTyCoInScopeSet :: [Type] -> [Coercion] -> InScopeSet

Generates an in-scope set from the free variables in a list of types and a list of coercions

Welll-scoped free variables

scopedSort :: [TyCoVar] -> [TyCoVar]

Do a topological sort on a list of tyvars, so that binders occur before occurrences E.g. given [ a::k, k::*, b::k ] it'll return a well-scoped list [ k::*, a::k, b::k ]

This is a deterministic sorting operation (that is, doesn't depend on Uniques).

It is also meant to be stable: that is, variables should not be reordered unnecessarily. This is specified in Note [ScopedSort] See also Note [Ordering of implicit variables] in RnTypes

tyCoVarsOfTypeWellScoped :: Type -> [TyVar]

Get the free vars of a type in scoped order

tyCoVarsOfTypesWellScoped :: [Type] -> [TyVar]

Get the free vars of types in scoped order