TyCon

Main TyCon data types

data TyCon

TyCons represent type constructors. Type constructors are introduced by things such as:

1) Data declarations: data Foo = ... creates the Foo type constructor of kind *

2) Type synonyms: type Foo = ... creates the Foo type constructor

3) Newtypes: newtype Foo a = MkFoo ... creates the Foo type constructor of kind * -> *

4) Class declarations: class Foo where creates the Foo type constructor of kind *

This data type also encodes a number of primitive, built in type constructors such as those for function and tuple types.

data AlgTyConRhs

Represents right-hand-sides of TyCons for algebraic types

Constructors

AbstractTyCon
Says that we know nothing about this data type, except that it's represented by a pointer. Used when we export a data type abstractly into an .hi file.

DataTyCon
Information about those TyCons derived from a data declaration. This includes data types with no constructors at all.

Fields

data_cons :: [DataCon]

The data type constructors; can be empty if the user declares the type to have no constructors

INVARIANT: Kept in order of increasing DataCon tag (see the tag assignment in DataCon.mkDataCon)

is_enum :: Bool

Cached value: is this an enumeration type? See Note [Enumeration types]

TupleTyCon

Fields

data_con :: DataCon

The unique constructor for the newtype. It has no existentials

tup_sort :: TupleSort

Is this a boxed, unboxed or constraint tuple?

SumTyCon

Fields

data_cons :: [DataCon]

The data type constructors; can be empty if the user declares the type to have no constructors

INVARIANT: Kept in order of increasing DataCon tag (see the tag assignment in DataCon.mkDataCon)

NewTyCon
Information about those TyCons derived from a newtype declaration

Fields

data_con :: DataCon

The unique constructor for the newtype. It has no existentials

nt_rhs :: Type

Cached value: the argument type of the constructor, which is just the representation type of the TyCon (remember that newtypes do not exist at runtime so need a different representation type).

The free TyVars of this type are the tyConTyVars from the corresponding TyCon

nt_etad_rhs :: ([TyVar], Type)

Same as the nt_rhs, but this time eta-reduced. Hence the list of TyVars in this field may be shorter than the declared arity of the TyCon.

nt_co :: CoAxiom Unbranched

visibleDataCons :: AlgTyConRhs -> [DataCon]

Both type classes as well as family instances imply implicit type constructors. These implicit type constructors refer to their parent structure (ie, the class or family from which they derive) using a type of the following form.

Extract those DataCons that we are able to learn about. Note that visibility in this sense does not correspond to visibility in the context of any particular user program!

data AlgTyConFlav

Constructors

VanillaAlgTyCon TyConRepName
An ordinary type constructor has no parent.

UnboxedAlgTyCon (Maybe TyConRepName)
An unboxed type constructor. The TyConRepName is a Maybe since we currently don't allow unboxed sums to be Typeable since there are too many of them. See #13276.

ClassTyCon Class TyConRepName
Type constructors representing a class dictionary. See Note [ATyCon for classes] in TyCoRep

DataFamInstTyCon (CoAxiom Unbranched) TyCon [Type]
Type constructors representing an *instance* of a *data* family. Parameters:

1) The type family in question

2) Instance types; free variables are the tyConTyVars of the current TyCon (not the family one). INVARIANT: the number of types matches the arity of the family TyCon

3) A CoTyCon identifying the representation type with the type instance family

isNoParent :: AlgTyConFlav -> Bool

data FamTyConFlav

Information pertaining to the expansion of a type synonym (type)

Constructors

DataFamilyTyCon TyConRepName
Represents an open type family without a fixed right hand side. Additional instances can appear at any time.

These are introduced by either a top level declaration:

data family T a :: *

Or an associated data type declaration, within a class declaration:

class C a b where
  data T b :: *

OpenSynFamilyTyCon
An open type synonym family e.g. type family F x y :: * -> *

ClosedSynFamilyTyCon (Maybe (CoAxiom Branched))
A closed type synonym family e.g. type family F x where { F Int = Bool }

AbstractClosedSynFamilyTyCon
A closed type synonym family declared in an hs-boot file with type family F a where ..

BuiltInSynFamTyCon BuiltInSynFamily
Built-in type family used by the TypeNats solver

data Role

Constructors

Nominal

Representational

Phantom

data Injectivity

Constructors

NotInjective

Injective [Bool]

data RuntimeRepInfo

Some promoted datacons signify extra info relevant to GHC. For example, the IntRep constructor of RuntimeRep corresponds to the IntRep constructor of PrimRep. This data structure allows us to store this information right in the TyCon. The other approach would be to look up things like RuntimeRep's PrimRep by known-key every time.

Constructors

NoRRI
an ordinary promoted data con

RuntimeRep ([Type] -> [PrimRep])
A constructor of RuntimeRep. The argument to the function should be the list of arguments to the promoted datacon.

VecCount Int
A constructor of VecCount

VecElem PrimElemRep
A constructor of VecElem

TyConBinder

type TyConBinder = TyVarBndr TyVar TyConBndrVis

data TyConBndrVis

Constructors

NamedTCB ArgFlag

AnonTCB

mkNamedTyConBinder :: ArgFlag -> TyVar -> TyConBinder

mkNamedTyConBinders :: ArgFlag -> [TyVar] -> [TyConBinder]

mkAnonTyConBinder :: TyVar -> TyConBinder

mkAnonTyConBinders :: [TyVar] -> [TyConBinder]

tyConBinderArgFlag :: TyConBinder -> ArgFlag

isNamedTyConBinder :: TyConBinder -> Bool

isVisibleTyConBinder :: TyVarBndr tv TyConBndrVis -> Bool

isInvisibleTyConBinder :: TyVarBndr tv TyConBndrVis -> Bool

Field labels

tyConFieldLabels :: TyCon -> [FieldLabel]

The labels for the fields of this particular TyCon

lookupTyConFieldLabel :: FieldLabelString -> TyCon -> Maybe FieldLabel

Look up a field label belonging to this TyCon

Constructing TyCons

mkAlgTyCon

Arguments

:: Name

-> [TyConBinder]
Binders of the TyCon

-> Kind
Result kind

-> [Role]
The roles for each TyVar

-> Maybe CType
The C type this type corresponds to when using the CAPI FFI

-> [PredType]
Stupid theta: see algTcStupidTheta

-> AlgTyConRhs
Information about data constructors

-> AlgTyConFlav
What flavour is it? (e.g. vanilla, type family)

-> Bool
Was the TyCon declared with GADT syntax?

-> TyCon

This is the making of an algebraic TyCon. Notably, you have to pass in the generic (in the -XGenerics sense) information about the type constructor - you can get hold of it easily (see Generics module)

mkClassTyCon :: Name -> [TyConBinder] -> [Role] -> AlgTyConRhs -> Class -> Name -> TyCon

Simpler specialization of mkAlgTyCon for classes

mkFunTyCon :: Name -> [TyConBinder] -> Name -> TyCon

Given the name of the function type constructor and it's kind, create the corresponding TyCon. It is recomended to use funTyCon if you want this functionality

mkPrimTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind, never levity-polymorphic

-> [Role]

-> TyCon

Create an unlifted primitive TyCon, such as Int#.

mkKindTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind

-> [Role]

-> Name

-> TyCon

Kind constructors

mkLiftedPrimTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind

-> [Role]

-> TyCon

Create a lifted primitive TyCon such as RealWorld

mkTupleTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
Result kind of the TyCon

-> Arity
Arity of the tuple TyCon

-> DataCon

-> TupleSort
Whether the tuple is boxed or unboxed

-> AlgTyConFlav

-> TyCon

mkSumTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
Kind of the resulting TyCon

-> Arity
Arity of the sum

-> [TyVar]
TyVars scoped over: see tyConTyVars

-> [DataCon]

-> AlgTyConFlav

-> TyCon

mkSynonymTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind

-> [Role]

-> Type

-> Bool

-> Bool

-> TyCon

Create a type synonym TyCon

mkFamilyTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind

-> Maybe Name

-> FamTyConFlav

-> Maybe Class

-> Injectivity

-> TyCon

Create a type family TyCon

mkPromotedDataCon :: DataCon -> Name -> TyConRepName -> [TyConBinder] -> Kind -> [Role] -> RuntimeRepInfo -> TyCon

Create a promoted data constructor TyCon Somewhat dodgily, we give it the same Name as the data constructor itself; when we pretty-print the TyCon we add a quote; see the Outputable TyCon instance

mkTcTyCon

Arguments

:: Name

-> [TyConBinder]

-> Kind
result kind only

-> Bool
Can this be unsaturated?

-> [TyVar]
Scoped type variables, see Note [TcTyCon]

-> TyCon

Makes a tycon suitable for use during type-checking. The only real need for this is for printing error messages during a recursive type/class type-checking knot. It has a kind because TcErrors sometimes calls typeKind. See also Note [Kind checking recursive type and class declarations] in TcTyClsDecls.

Predicates on TyCons

isAlgTyCon :: TyCon -> Bool

Returns True if the supplied TyCon resulted from either a data or newtype declaration

isVanillaAlgTyCon :: TyCon -> Bool

Returns True for vanilla AlgTyCons -- that is, those created with a data or newtype declaration.

isClassTyCon :: TyCon -> Bool

Is this TyCon that for a class instance?

isFamInstTyCon :: TyCon -> Bool

Is this TyCon that for a data family instance?

isFunTyCon :: TyCon -> Bool

isPrimTyCon :: TyCon -> Bool

Does this TyCon represent something that cannot be defined in Haskell?

isTupleTyCon :: TyCon -> Bool

Does this TyCon represent a tuple?

NB: when compiling Data.Tuple, the tycons won't reply True to isTupleTyCon, because they are built as AlgTyCons. However they get spat into the interface file as tuple tycons, so I don't think it matters.

isUnboxedTupleTyCon :: TyCon -> Bool

Is this the TyCon for an unboxed tuple?

isBoxedTupleTyCon :: TyCon -> Bool

Is this the TyCon for a boxed tuple?

isUnboxedSumTyCon :: TyCon -> Bool

Is this the TyCon for an unboxed sum?

isPromotedTupleTyCon :: TyCon -> Bool

Is this the TyCon for a promoted tuple?

isTypeSynonymTyCon :: TyCon -> Bool

Is this a TyCon representing a regular H98 type synonym (type)?

mightBeUnsaturatedTyCon :: TyCon -> Bool

True iff we can decompose (T a b c) into ((T a b) c) I.e. is it injective and generative w.r.t nominal equality? That is, if (T a b) ~N d e f, is it always the case that (T ~N d), (a ~N e) and (b ~N f)? Specifically NOT true of synonyms (open and otherwise)

It'd be unusual to call mightBeUnsaturatedTyCon on a regular H98 type synonym, because you should probably have expanded it first But regardless, it's not decomposable

isPromotedDataCon :: TyCon -> Bool

Is this a PromotedDataCon?

isPromotedDataCon_maybe :: TyCon -> Maybe DataCon

Retrieves the promoted DataCon if this is a PromotedDataCon;

isKindTyCon :: TyCon -> Bool

Is this tycon really meant for use at the kind level? That is, should it be permitted without -XDataKinds?

isLiftedTypeKindTyConName :: Name -> Bool

isTauTyCon :: TyCon -> Bool

isFamFreeTyCon :: TyCon -> Bool

isDataTyCon :: TyCon -> Bool

Returns True for data types that are definitely represented by heap-allocated constructors. These are scrutinised by Core-level case expressions, and they get info tables allocated for them.

Generally, the function will be true for all data types and false for newtypes, unboxed tuples, unboxed sums and type family TyCons. But it is not guaranteed to return True in all cases that it could.

NB: for a data type family, only the instance TyCons get an info table. The family declaration TyCon does not

isProductTyCon :: TyCon -> Bool

isDataProductTyCon_maybe :: TyCon -> Maybe DataCon

isDataSumTyCon_maybe :: TyCon -> Maybe [DataCon]

isEnumerationTyCon :: TyCon -> Bool

Is this an algebraic TyCon which is just an enumeration of values?

isNewTyCon :: TyCon -> Bool

Is this TyCon that for a newtype

isAbstractTyCon :: TyCon -> Bool

Test if the TyCon is algebraic but abstract (invisible data constructors)

isFamilyTyCon :: TyCon -> Bool

Is this a TyCon, synonym or otherwise, that defines a family?

isOpenFamilyTyCon :: TyCon -> Bool

Is this a TyCon, synonym or otherwise, that defines a family with instances?

isTypeFamilyTyCon :: TyCon -> Bool

Is this a synonym TyCon that can have may have further instances appear?

isDataFamilyTyCon :: TyCon -> Bool

Is this a synonym TyCon that can have may have further instances appear?

isOpenTypeFamilyTyCon :: TyCon -> Bool

Is this an open type family TyCon?

isClosedSynFamilyTyConWithAxiom_maybe :: TyCon -> Maybe (CoAxiom Branched)

Is this a non-empty closed type family? Returns Nothing for abstract or empty closed families.

familyTyConInjectivityInfo :: TyCon -> Injectivity

Try to read the injectivity information from a FamilyTyCon. For every other TyCon this function panics.

isBuiltInSynFamTyCon_maybe :: TyCon -> Maybe BuiltInSynFamily

isUnliftedTyCon :: TyCon -> Bool

Is this TyCon unlifted (i.e. cannot contain bottom)? Note that this can only be true for primitive and unboxed-tuple TyCons

isGadtSyntaxTyCon :: TyCon -> Bool

Is this an algebraic TyCon declared with the GADT syntax?

isInjectiveTyCon :: TyCon -> Role -> Bool

isInjectiveTyCon is true of TyCons for which this property holds (where X is the role passed in): If (T a1 b1 c1) ~X (T a2 b2 c2), then (a1 ~X1 a2), (b1 ~X2 b2), and (c1 ~X3 c2) (where X1, X2, and X3, are the roles given by tyConRolesX tc X) See also Note [Decomposing equality] in TcCanonical

isGenerativeTyCon :: TyCon -> Role -> Bool

isGenerativeTyCon is true of TyCons for which this property holds (where X is the role passed in): If (T tys ~X t), then (t's head ~X T). See also Note [Decomposing equality] in TcCanonical

isGenInjAlgRhs :: AlgTyConRhs -> Bool

Is this an AlgTyConRhs of a TyCon that is generative and injective with respect to representational equality?

isTyConAssoc :: TyCon -> Bool

Are we able to extract information TyVar to class argument list mapping from a given TyCon?

tyConAssoc_maybe :: TyCon -> Maybe Class

isImplicitTyCon :: TyCon -> Bool

Identifies implicit tycons that, in particular, do not go into interface files (because they are implicitly reconstructed when the interface is read).

Note that:

Associated families are implicit, as they are re-constructed from the class declaration in which they reside, and

Family instances are not implicit as they represent the instance body (similar to a dfun does that for a class instance).

Tuples are implicit iff they have a wired-in name (namely: boxed and unboxed tupeles are wired-in and implicit, but constraint tuples are not)

isTyConWithSrcDataCons :: TyCon -> Bool

Check if the tycon actually refers to a proper `data` or `newtype` with user defined constructors rather than one from a class or other construction.

isTcTyCon :: TyCon -> Bool

Is this a TcTyCon? (That is, one only used during type-checking?)

isTcLevPoly :: TyCon -> Bool

Could this TyCon ever be levity-polymorphic when fully applied? True is safe. False means we're sure. Does only a quick check based on the TyCon's category. Precondition: The fully-applied TyCon has kind (TYPE blah)

Extracting information out of TyCons

tyConName :: TyCon -> Name

Name of the constructor

tyConSkolem :: TyCon -> Bool

Returns whether or not this TyCon is definite, or a hole that may be filled in at some later point. See Note [Skolem abstract data]

tyConKind :: TyCon -> Kind

Kind of this TyCon

tyConUnique :: TyCon -> Unique

A Unique of this TyCon. Invariant: identical to Unique of Name stored in tyConName field.

tyConTyVars :: TyCon -> [TyVar]

TyVar binders

tyConCType :: TyCon -> Maybe CType

The C type that should be used for this type when using the FFI and CAPI

tyConCType_maybe :: TyCon -> Maybe CType

tyConDataCons :: TyCon -> [DataCon]

As tyConDataCons_maybe, but returns the empty list of constructors if no constructors could be found

tyConDataCons_maybe :: TyCon -> Maybe [DataCon]

Determine the DataCons originating from the given TyCon, if the TyCon is the sort that can have any constructors (note: this does not include abstract algebraic types)

tyConSingleDataCon_maybe :: TyCon -> Maybe DataCon

If the given TyCon has a single data constructor, i.e. it is a data type with one alternative, a tuple type or a newtype then that constructor is returned. If the TyCon has more than one constructor, or represents a primitive or function type constructor then Nothing is returned. In any other case, the function panics

tyConSingleDataCon :: TyCon -> DataCon

tyConSingleAlgDataCon_maybe :: TyCon -> Maybe DataCon

tyConFamilySize :: TyCon -> Int

Determine the number of value constructors a TyCon has. Panics if the TyCon is not algebraic or a tuple

tyConStupidTheta :: TyCon -> [PredType]

Find the "stupid theta" of the TyCon. A "stupid theta" is the context to the left of an algebraic type declaration, e.g. Eq a in the declaration data Eq a => T a ...

tyConArity :: TyCon -> Arity

Arity

tyConRoles :: TyCon -> [Role]

Get the list of roles for the type parameters of a TyCon

tyConFlavour :: TyCon -> String

tyConTuple_maybe :: TyCon -> Maybe TupleSort

tyConClass_maybe :: TyCon -> Maybe Class

If this TyCon is that for a class instance, return the class it is for. Otherwise returns Nothing

tyConATs :: TyCon -> [TyCon]

Return the associated types of the TyCon, if any

tyConFamInst_maybe :: TyCon -> Maybe (TyCon, [Type])

If this TyCon is that of a data family instance, return the family in question and the instance types. Otherwise, return Nothing

tyConFamInstSig_maybe :: TyCon -> Maybe (TyCon, [Type], CoAxiom Unbranched)

tyConFamilyCoercion_maybe :: TyCon -> Maybe (CoAxiom Unbranched)

If this TyCon is that of a data family instance, return a TyCon which represents a coercion identifying the representation type with the type instance family. Otherwise, return Nothing

tyConFamilyResVar_maybe :: TyCon -> Maybe Name

Extract type variable naming the result of injective type family

synTyConDefn_maybe :: TyCon -> Maybe ([TyVar], Type)

Extract the TyVars bound by a vanilla type synonym and the corresponding (unsubstituted) right hand side.

synTyConRhs_maybe :: TyCon -> Maybe Type

Extract the information pertaining to the right hand side of a type synonym (type) declaration.

famTyConFlav_maybe :: TyCon -> Maybe FamTyConFlav

Extract the flavour of a type family (with all the extra information that it carries)

famTcResVar :: TyCon -> Maybe Name

Name of result type variable, used for pretty-printing with --show-iface and for reifying TyCon in Template Haskell

algTyConRhs :: TyCon -> AlgTyConRhs

Extract an AlgTyConRhs with information about data constructors from an algebraic or tuple TyCon. Panics for any other sort of TyCon

newTyConRhs :: TyCon -> ([TyVar], Type)

Extract the bound type variables and type expansion of a type synonym TyCon. Panics if the TyCon is not a synonym

newTyConEtadArity :: TyCon -> Int

The number of type parameters that need to be passed to a newtype to resolve it. May be less than in the definition if it can be eta-contracted.

newTyConEtadRhs :: TyCon -> ([TyVar], Type)

Extract the bound type variables and type expansion of an eta-contracted type synonym TyCon. Panics if the TyCon is not a synonym

unwrapNewTyCon_maybe :: TyCon -> Maybe ([TyVar], Type, CoAxiom Unbranched)

Take a TyCon apart into the TyVars it scopes over, the Type it expands into, and (possibly) a coercion from the representation type to the newtype. Returns Nothing if this is not possible.

unwrapNewTyConEtad_maybe :: TyCon -> Maybe ([TyVar], Type, CoAxiom Unbranched)

newTyConDataCon_maybe :: TyCon -> Maybe DataCon

algTcFields :: TyCon -> FieldLabelEnv

Maps a label to information about the field

tyConRuntimeRepInfo :: TyCon -> RuntimeRepInfo

Extract any RuntimeRepInfo from this TyCon

tyConBinders :: TyCon -> [TyConBinder]

Full binders

tyConResKind :: TyCon -> Kind

Result kind

tcTyConScopedTyVars :: TyCon -> [TyVar]

Scoped tyvars over the tycon's body. See Note [TcTyCon]

Manipulating TyCons

expandSynTyCon_maybe

Arguments

:: TyCon

-> [tyco]
Arguments to TyCon

-> Maybe ([(TyVar, tyco)], Type, [tyco])
Returns a TyVar substitution, the body type of the synonym (not yet substituted) and any arguments remaining from the application

Expand a type synonym application, if any

makeRecoveryTyCon :: TyCon -> TyCon

Make an fake, recovery TyCon from an existing one. Used when recovering from errors

newTyConCo :: TyCon -> CoAxiom Unbranched

newTyConCo_maybe :: TyCon -> Maybe (CoAxiom Unbranched)

Extracts the newtype coercion from such a TyCon, which can be used to construct something with the newtypes type from its representation type (right hand side). If the supplied TyCon is not a newtype, returns Nothing

pprPromotionQuote :: TyCon -> SDoc

mkTyConKind :: [TyConBinder] -> Kind -> Kind

Runtime type representation

type TyConRepName = Name

tyConRepName_maybe :: TyCon -> Maybe TyConRepName

mkPrelTyConRepName :: Name -> TyConRepName

Make a Name for the Typeable representation of the given wired-in type

tyConRepModOcc :: Module -> OccName -> (Module, OccName)

The name (and defining module) for the Typeable representation (TyCon) of a type constructor.

See Note [Grand plan for Typeable] in TcTypeable in TcTypeable.

Primitive representations of Types

data PrimRep

A PrimRep is an abstraction of a type. It contains information that the code generator needs in order to pass arguments, return results, and store values of this type.

Constructors

VoidRep

LiftedRep

UnliftedRep
Unlifted pointer

IntRep
Signed, word-sized value

WordRep
Unsigned, word-sized value

Int64Rep
Signed, 64 bit value (with 32-bit words only)

Word64Rep
Unsigned, 64 bit value (with 32-bit words only)

AddrRep
A pointer, but not to a Haskell value (use '(Un)liftedRep')

FloatRep

DoubleRep

VecRep Int PrimElemRep
A vector

data PrimElemRep

Constructors

Int8ElemRep

Int16ElemRep

Int32ElemRep

Int64ElemRep

Word8ElemRep

Word16ElemRep

Word32ElemRep

Word64ElemRep

FloatElemRep

DoubleElemRep

isVoidRep :: PrimRep -> Bool

isGcPtrRep :: PrimRep -> Bool

primRepSizeW :: DynFlags -> PrimRep -> Int

Find the size of a PrimRep, in words

primElemRepSizeB :: PrimElemRep -> Int

primRepIsFloat :: PrimRep -> Maybe Bool

Return if Rep stands for floating type, returns Nothing for vector types.

Recursion breaking

data RecTcChecker

initRecTc :: RecTcChecker

checkRecTc :: RecTcChecker -> TyCon -> Maybe RecTcChecker