TysWiredIn

Description

This module is about types that can be defined in Haskell, but which must be wired into the compiler nonetheless. C.f module TysPrim

Helper functions defined here

mkWiredInTyConName :: BuiltInSyntax -> Module -> FastString -> Unique -> TyCon -> Name

mkWiredInIdName :: Module -> FastString -> Unique -> Id -> Name

mkFunKind :: Kind -> Kind -> Kind

mkForAllKind :: TyVar -> ArgFlag -> Kind -> Kind

All wired in things

wiredInTyCons :: [TyCon]

isBuiltInOcc_maybe :: OccName -> Maybe Name

Built-in syntax isn't "in scope" so these OccNames map to wired-in Names with BuiltInSyntax. However, this should only be necessary while resolving names produced by Template Haskell splices since we take care to encode built-in syntax names specially in interface files. See Note [Symbol table representation of names].

Moreover, there is no need to include names of things that the user can't write (e.g. type representation bindings like $tc(,,,)).

Bool

boolTy :: Type

boolTyCon :: TyCon

boolTyCon_RDR :: RdrName

boolTyConName :: Name

trueDataCon :: DataCon

trueDataConId :: Id

true_RDR :: RdrName

falseDataCon :: DataCon

falseDataConId :: Id

false_RDR :: RdrName

promotedFalseDataCon :: TyCon

promotedTrueDataCon :: TyCon

Ordering

orderingTyCon :: TyCon

ltDataCon :: DataCon

ltDataConId :: Id

eqDataCon :: DataCon

eqDataConId :: Id

gtDataCon :: DataCon

gtDataConId :: Id

promotedLTDataCon :: TyCon

promotedEQDataCon :: TyCon

promotedGTDataCon :: TyCon

Boxing primitive types

boxingDataCon_maybe :: TyCon -> Maybe DataCon

Char

charTyCon :: TyCon

charDataCon :: DataCon

charTyCon_RDR :: RdrName

charTy :: Type

stringTy :: Type

charTyConName :: Name

Double

doubleTyCon :: TyCon

doubleDataCon :: DataCon

doubleTy :: Type

doubleTyConName :: Name

Float

floatTyCon :: TyCon

floatDataCon :: DataCon

floatTy :: Type

floatTyConName :: Name

Int

intTyCon :: TyCon

intDataCon :: DataCon

intTyCon_RDR :: RdrName

intDataCon_RDR :: RdrName

intTyConName :: Name

intTy :: Type

Word

wordTyCon :: TyCon

wordDataCon :: DataCon

wordTyConName :: Name

wordTy :: Type

Word8

word8TyCon :: TyCon

word8DataCon :: DataCon

word8TyConName :: Name

word8Ty :: Type

List

listTyCon :: TyCon

listTyCon_RDR :: RdrName

listTyConName :: Name

listTyConKey :: Unique

nilDataCon :: DataCon

nilDataConName :: Name

nilDataConKey :: Unique

consDataCon_RDR :: RdrName

consDataCon :: DataCon

consDataConName :: Name

promotedNilDataCon :: TyCon

promotedConsDataCon :: TyCon

mkListTy :: Type -> Type

mkPromotedListTy

Arguments

:: Kind
of the elements of the list

-> [Type]
elements

-> Type

Make a *promoted* list.

Maybe

maybeTyCon :: TyCon

maybeTyConName :: Name

nothingDataCon :: DataCon

nothingDataConName :: Name

promotedNothingDataCon :: TyCon

justDataCon :: DataCon

justDataConName :: Name

promotedJustDataCon :: TyCon

Tuples

mkTupleTy :: Boxity -> [Type] -> Type

Make a tuple type. The list of types should not include any RuntimeRep specifications.

mkBoxedTupleTy :: [Type] -> Type

Build the type of a small tuple that holds the specified type of thing

tupleTyCon :: Boxity -> Arity -> TyCon

tupleDataCon :: Boxity -> Arity -> DataCon

tupleTyConName :: TupleSort -> Arity -> Name

promotedTupleDataCon :: Boxity -> Arity -> TyCon

unitTyCon :: TyCon

unitDataCon :: DataCon

unitDataConId :: Id

unitTy :: Type

unitTyConKey :: Unique

pairTyCon :: TyCon

unboxedUnitTyCon :: TyCon

unboxedUnitDataCon :: DataCon

unboxedTupleKind :: [Type] -> Kind

Specialization of unboxedTupleSumKind for tuples

unboxedSumKind :: [Type] -> Kind

Specialization of unboxedTupleSumKind for sums

Constraint tuples

cTupleTyConName :: Arity -> Name

cTupleTyConNames :: [Name]

isCTupleTyConName :: Name -> Bool

cTupleDataConName :: Arity -> Name

cTupleDataConNames :: [Name]

Any

anyTyCon :: TyCon

anyTy :: Type

anyTypeOfKind :: Kind -> Type

Sums

mkSumTy :: [Type] -> Type

sumTyCon :: Arity -> TyCon

Type constructor for n-ary unboxed sum.

sumDataCon :: ConTag -> Arity -> DataCon

Data constructor for i-th alternative of a n-ary unboxed sum.

Kinds

typeNatKindCon :: TyCon

typeNatKind :: Kind

typeSymbolKindCon :: TyCon

typeSymbolKind :: Kind

isLiftedTypeKindTyConName :: Name -> Bool

liftedTypeKind :: Kind

constraintKind :: Kind

starKindTyCon :: TyCon

starKindTyConName :: Name

unicodeStarKindTyCon :: TyCon

unicodeStarKindTyConName :: Name

liftedTypeKindTyCon :: TyCon

constraintKindTyCon :: TyCon

Parallel arrays

mkPArrTy :: Type -> Type

Construct a type representing the application of the parallel array constructor

parrTyCon :: TyCon

Represents the type constructor of parallel arrays

This must match the definition in PrelPArr

NB: Although the constructor is given here, it will not be accessible in user code as it is not in the environment of any compiled module except PrelPArr.

parrFakeCon :: Arity -> DataCon

Fake array constructors

These constructors are never really used to represent array values; however, they are very convenient during desugaring (and, in particular, in the pattern matching compiler) to treat array pattern just like yet another constructor pattern

isPArrTyCon :: TyCon -> Bool

Check whether a type constructor is the constructor for parallel arrays

isPArrFakeCon :: DataCon -> Bool

Checks whether a data constructor is a fake constructor for parallel arrays

parrTyCon_RDR :: RdrName

parrTyConName :: Name

Equality predicates

heqTyCon :: TyCon

heqClass :: Class

heqDataCon :: DataCon

coercibleTyCon :: TyCon

coercibleDataCon :: DataCon

coercibleClass :: Class

RuntimeRep and friends

runtimeRepTyCon :: TyCon

vecCountTyCon :: TyCon

vecElemTyCon :: TyCon

runtimeRepTy :: Type

liftedRepTy :: Type

liftedRepDataCon :: DataCon

liftedRepDataConTyCon :: TyCon

vecRepDataConTyCon :: TyCon

tupleRepDataConTyCon :: TyCon

sumRepDataConTyCon :: TyCon

liftedRepDataConTy :: Type

unliftedRepDataConTy :: Type

intRepDataConTy :: Type

wordRepDataConTy :: Type

int64RepDataConTy :: Type

word64RepDataConTy :: Type

addrRepDataConTy :: Type

floatRepDataConTy :: Type

doubleRepDataConTy :: Type

vec2DataConTy :: Type

vec4DataConTy :: Type

vec8DataConTy :: Type

vec16DataConTy :: Type

vec32DataConTy :: Type

vec64DataConTy :: Type

int8ElemRepDataConTy :: Type

int16ElemRepDataConTy :: Type

int32ElemRepDataConTy :: Type

int64ElemRepDataConTy :: Type

word8ElemRepDataConTy :: Type

word16ElemRepDataConTy :: Type

word32ElemRepDataConTy :: Type

word64ElemRepDataConTy :: Type

floatElemRepDataConTy :: Type

doubleElemRepDataConTy :: Type