UniqDFM

Unique-keyed deterministic mappings

data UniqDFM ele

Type of unique deterministic finite maps

Manipulating those mappings

emptyUDFM :: UniqDFM elt

unitUDFM :: Uniquable key => key -> elt -> UniqDFM elt

addToUDFM :: Uniquable key => UniqDFM elt -> key -> elt -> UniqDFM elt

addToUDFM_C :: Uniquable key => (elt -> elt -> elt) -> UniqDFM elt -> key -> elt -> UniqDFM elt

addListToUDFM :: Uniquable key => UniqDFM elt -> [(key, elt)] -> UniqDFM elt

delFromUDFM :: Uniquable key => UniqDFM elt -> key -> UniqDFM elt

delListFromUDFM :: Uniquable key => UniqDFM elt -> [key] -> UniqDFM elt

Delete a list of elements from a UniqDFM

adjustUDFM :: Uniquable key => (elt -> elt) -> UniqDFM elt -> key -> UniqDFM elt

Apply a function to a particular element

alterUDFM :: Uniquable key => (Maybe elt -> Maybe elt) -> UniqDFM elt -> key -> UniqDFM elt

The expression (alterUDFM f k map) alters value x at k, or absence thereof. alterUDFM can be used to insert, delete, or update a value in UniqDFM. Use addToUDFM, delFromUDFM or adjustUDFM when possible, they are more efficient.

mapUDFM :: (elt1 -> elt2) -> UniqDFM elt1 -> UniqDFM elt2

Map a function over every value in a UniqDFM

plusUDFM :: UniqDFM elt -> UniqDFM elt -> UniqDFM elt

plusUDFM_C :: (elt -> elt -> elt) -> UniqDFM elt -> UniqDFM elt -> UniqDFM elt

lookupUDFM :: Uniquable key => UniqDFM elt -> key -> Maybe elt

lookupUDFM_Directly :: UniqDFM elt -> Unique -> Maybe elt

elemUDFM :: Uniquable key => key -> UniqDFM elt -> Bool

foldUDFM :: (elt -> a -> a) -> a -> UniqDFM elt -> a

Performs a deterministic fold over the UniqDFM. It's O(n log n) while the corresponding function on UniqFM is O(n).

eltsUDFM :: UniqDFM elt -> [elt]

filterUDFM :: (elt -> Bool) -> UniqDFM elt -> UniqDFM elt

filterUDFM_Directly :: (Unique -> elt -> Bool) -> UniqDFM elt -> UniqDFM elt

isNullUDFM :: UniqDFM elt -> Bool

sizeUDFM :: UniqDFM elt -> Int

intersectUDFM :: UniqDFM elt -> UniqDFM elt -> UniqDFM elt

udfmIntersectUFM :: UniqDFM elt1 -> UniqFM elt2 -> UniqDFM elt1

intersectsUDFM :: UniqDFM elt -> UniqDFM elt -> Bool

disjointUDFM :: UniqDFM elt -> UniqDFM elt -> Bool

disjointUdfmUfm :: UniqDFM elt -> UniqFM elt2 -> Bool

equalKeysUDFM :: UniqDFM a -> UniqDFM b -> Bool

minusUDFM :: UniqDFM elt1 -> UniqDFM elt2 -> UniqDFM elt1

listToUDFM :: Uniquable key => [(key, elt)] -> UniqDFM elt

udfmMinusUFM :: UniqDFM elt1 -> UniqFM elt2 -> UniqDFM elt1

partitionUDFM :: (elt -> Bool) -> UniqDFM elt -> (UniqDFM elt, UniqDFM elt)

Partition UniqDFM into two UniqDFMs according to the predicate

anyUDFM :: (elt -> Bool) -> UniqDFM elt -> Bool

allUDFM :: (elt -> Bool) -> UniqDFM elt -> Bool

pprUniqDFM :: (a -> SDoc) -> UniqDFM a -> SDoc

pprUDFM

Arguments

:: UniqDFM a
The things to be pretty printed

-> ([a] -> SDoc)
The pretty printing function to use on the elements

-> SDoc
SDoc where the things have been pretty printed

udfmToList :: UniqDFM elt -> [(Unique, elt)]

Converts UniqDFM to a list, with elements in deterministic order. It's O(n log n) while the corresponding function on UniqFM is O(n).

udfmToUfm :: UniqDFM elt -> UniqFM elt

This allows for lossy conversion from UniqDFM to UniqFM

nonDetFoldUDFM :: (elt -> a -> a) -> a -> UniqDFM elt -> a

Performs a nondeterministic fold over the UniqDFM. It's O(n), same as the corresponding function on UniqFM. If you use this please provide a justification why it doesn't introduce nondeterminism.

alwaysUnsafeUfmToUdfm :: UniqFM elt -> UniqDFM elt