UniqSet

Unique set type

data UniqSet a

getUniqSet :: UniqSet a -> UniqFM a

pprUniqSet :: (a -> SDoc) -> UniqSet a -> SDoc

Manipulating these sets

emptyUniqSet :: UniqSet a

unitUniqSet :: Uniquable a => a -> UniqSet a

mkUniqSet :: Uniquable a => [a] -> UniqSet a

addOneToUniqSet :: Uniquable a => UniqSet a -> a -> UniqSet a

addListToUniqSet :: Uniquable a => UniqSet a -> [a] -> UniqSet a

delOneFromUniqSet :: Uniquable a => UniqSet a -> a -> UniqSet a

delOneFromUniqSet_Directly :: UniqSet a -> Unique -> UniqSet a

delListFromUniqSet :: Uniquable a => UniqSet a -> [a] -> UniqSet a

delListFromUniqSet_Directly :: UniqSet a -> [Unique] -> UniqSet a

unionUniqSets :: UniqSet a -> UniqSet a -> UniqSet a

unionManyUniqSets :: [UniqSet a] -> UniqSet a

minusUniqSet :: UniqSet a -> UniqSet a -> UniqSet a

uniqSetMinusUFM :: UniqSet a -> UniqFM b -> UniqSet a

intersectUniqSets :: UniqSet a -> UniqSet a -> UniqSet a

restrictUniqSetToUFM :: UniqSet a -> UniqFM b -> UniqSet a

uniqSetAny :: (a -> Bool) -> UniqSet a -> Bool

uniqSetAll :: (a -> Bool) -> UniqSet a -> Bool

elementOfUniqSet :: Uniquable a => a -> UniqSet a -> Bool

elemUniqSet_Directly :: Unique -> UniqSet a -> Bool

filterUniqSet :: (a -> Bool) -> UniqSet a -> UniqSet a

filterUniqSet_Directly :: (Unique -> elt -> Bool) -> UniqSet elt -> UniqSet elt

sizeUniqSet :: UniqSet a -> Int

isEmptyUniqSet :: UniqSet a -> Bool

lookupUniqSet :: Uniquable a => UniqSet b -> a -> Maybe b

lookupUniqSet_Directly :: UniqSet a -> Unique -> Maybe a

partitionUniqSet :: (a -> Bool) -> UniqSet a -> (UniqSet a, UniqSet a)

mapUniqSet :: Uniquable b => (a -> b) -> UniqSet a -> UniqSet b

unsafeUFMToUniqSet :: UniqFM a -> UniqSet a

unsafeUFMToUniqSet converts a UniqFM a into a UniqSet a assuming, without checking, that it maps each Unique to a value that has that Unique. See Note [Unsound mapUniqSet].

nonDetEltsUniqSet :: UniqSet elt -> [elt]

nonDetKeysUniqSet :: UniqSet elt -> [Unique]

nonDetFoldUniqSet :: (elt -> a -> a) -> a -> UniqSet elt -> a

nonDetFoldUniqSet_Directly :: (Unique -> elt -> a -> a) -> a -> UniqSet elt -> a