Var

Description

GHC uses several kinds of name internally:

OccName: see OccName

RdrName: see RdrName

Name: see Name

Id: see Id

Var is a synonym for the Id type but it may additionally potentially contain type variables, which have a Kind rather than a Type and only contain some extra details during typechecking.

These Var names may either be global or local, see Var

Global Ids and Vars are those that are imported or correspond to a data constructor, primitive operation, or record selectors. Local Ids and Vars are those bound within an expression (e.g. by a lambda) or at the top level of the module being compiled.

The main data type and synonyms

data Var

Variable

Essentially a typed Name, that may also contain some additional information about the Var and it's use sites.

type CoVar = Id

Coercion Variable

type Id = Var

Identifier

type NcId = Id

type DictId = EvId

Dictionary Identifier

type DFunId = Id

Dictionary Function Identifier

type EvVar = EvId

Evidence Variable

type EqVar = EvId

Equality Variable

type EvId = Id

Evidence Identifier

type IpId = EvId

Implicit parameter Identifier

type JoinId = Id

type TyVar = Var

Type or kind Variable

type TypeVar = Var

Type Variable

type KindVar = Var

Kind Variable

type TKVar = Var

Type or Kind Variable

type TyCoVar = Id

Type or Coercion Variable

In and Out variants

type InVar = Var

type InCoVar = CoVar

type InId = Id

type InTyVar = TyVar

type OutVar = Var

type OutCoVar = CoVar

type OutId = Id

type OutTyVar = TyVar

Taking Vars apart

varName :: Var -> Name

varUnique :: Var -> Unique

varType :: Var -> Kind

The type or kind of the Var in question

Modifying Vars

setVarName :: Var -> Name -> Var

setVarUnique :: Var -> Unique -> Var

setVarType :: Id -> Type -> Id

updateVarType :: (Type -> Type) -> Id -> Id

updateVarTypeM :: Monad m => (Type -> m Type) -> Id -> m Id

Constructing, taking apart, modifying Ids

mkGlobalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

mkLocalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

mkExportedLocalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

Exported Vars will not be removed as dead code

mkCoVar :: Name -> Type -> CoVar

idInfo :: HasDebugCallStack => Id -> IdInfo

idDetails :: Id -> IdDetails

lazySetIdInfo :: Id -> IdInfo -> Var

setIdDetails :: Id -> IdDetails -> Id

globaliseId :: Id -> Id

If it's a local, make it global

setIdExported :: Id -> Id

Exports the given local Id. Can also be called on global Ids, such as data constructors and class operations, which are born as global Ids and automatically exported

setIdNotExported :: Id -> Id

We can only do this to LocalIds

Predicates

isId :: Var -> Bool

isTyVar :: Var -> Bool

isTcTyVar :: Var -> Bool

isLocalVar :: Var -> Bool

isLocalVar returns True for type variables as well as local Ids These are the variables that we need to pay attention to when finding free variables, or doing dependency analysis.

isLocalId :: Var -> Bool

isCoVar :: Var -> Bool

isNonCoVarId :: Var -> Bool

isTyCoVar :: Var -> Bool

isGlobalId :: Var -> Bool

isExportedId :: Var -> Bool

isExportedIdVar means "don't throw this away"

mustHaveLocalBinding :: Var -> Bool

mustHaveLocalBinding returns True of Ids and TyVars that must have a binding in this module. The converse is not quite right: there are some global Ids that must have bindings, such as record selectors. But that doesn't matter, because it's only used for assertions

TyVar's

data TyVarBndr tyvar argf

Constructors

TvBndr tyvar argf

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [TyBinders and ArgFlags] in TyCoRep

Constructors

Required

Specified

Inferred

type TyVarBinder = TyVarBndr TyVar ArgFlag

Type Variable Binder

A TyVarBinder is the binder of a ForAllTy It's convenient to define this synonym here rather its natural home in TyCoRep, because it's used in DataCon.hs-boot

binderVar :: TyVarBndr tv argf -> tv

binderVars :: [TyVarBndr tv argf] -> [tv]

binderArgFlag :: TyVarBndr tv argf -> argf

binderKind :: TyVarBndr TyVar argf -> Kind

isVisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is written in Haskell?

isInvisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is not written in Haskell?

sameVis :: ArgFlag -> ArgFlag -> Bool

Do these denote the same level of visibility? Required arguments are visible, others are not. So this function equates Specified and Inferred. Used for printing.

Constructing TyVar's

mkTyVar :: Name -> Kind -> TyVar

mkTcTyVar :: Name -> Kind -> TcTyVarDetails -> TyVar

Taking TyVars apart

tyVarName :: TyVar -> Name

tyVarKind :: TyVar -> Kind

tcTyVarDetails :: TyVar -> TcTyVarDetails

setTcTyVarDetails :: TyVar -> TcTyVarDetails -> TyVar

Modifying TyVars

setTyVarName :: TyVar -> Name -> TyVar

setTyVarUnique :: TyVar -> Unique -> TyVar

setTyVarKind :: TyVar -> Kind -> TyVar

updateTyVarKind :: (Kind -> Kind) -> TyVar -> TyVar

updateTyVarKindM :: Monad m => (Kind -> m Kind) -> TyVar -> m TyVar

nonDetCmpVar :: Var -> Var -> Ordering

Compare Vars by their Uniques. This is what Ord Var does, provided here to make it explicit at the call-site that it can introduce non-determinism. See Note [Unique Determinism]