Vectorise.Builtins.Initialise

Initialisation

initBuiltins :: DsM Builtins

Create the initial map of builtin types and functions.

initBuiltinVars :: Builtins -> DsM [(Var, Var)]

Get the mapping of names in the Prelude to names in the DPH library.