Vectorise.Generic.PAMethods

Description

Generate methods for the PA class.

TODO: there is a large amount of redundancy here between the a, PData a, and PDatas a forms. See if we can factor some of this out.

Documentation

buildPReprTyCon :: TyCon -> TyCon -> SumRepr -> VM FamInst

buildPAScAndMethods :: VM [(String, PAInstanceBuilder)]