Vectorise.Var

Description

Vectorise variables and literals.

Documentation

vectBndr :: Var -> VM VVar

Vectorise a binder variable, along with its attached type.

vectBndrNew :: Var -> FastString -> VM VVar

Vectorise a binder variable, along with its attached type, but give the result a new name.

vectBndrIn :: Var -> VM a -> VM (VVar, a)

Vectorise a binder then run a computation with that binder in scope.

vectBndrNewIn :: Var -> FastString -> VM a -> VM (VVar, a)

Vectorise a binder, give it a new name, then run a computation with that binder in scope.

vectBndrsIn :: [Var] -> VM a -> VM ([VVar], a)

Vectorise some binders, then run a computation with them in scope.

vectVar :: Var -> VM VExpr

Vectorise a variable, producing the vectorised and lifted versions.

vectConst :: CoreExpr -> VM VExpr

Constants are lifted by replication along the integer context in the VM state for the number of elements in the result array.