X86.Cond

Documentation

data Cond

Constructors

ALWAYS

EQQ

GE

GEU

GTT

GU

LE

LEU

LTT

LU

NE

NEG

POS

CARRY

OFLO

PARITY

NOTPARITY

condUnsigned :: Cond -> Bool

condToSigned :: Cond -> Cond

condToUnsigned :: Cond -> Cond

maybeFlipCond :: Cond -> Maybe Cond

maybeFlipCond c returns Just c' if it is possible to flip the arguments to the conditional c, and the new condition should be c'.