AsmCodeGen

Module entry point

nativeCodeGen :: DynFlags -> Module -> ModLocation -> Handle -> UniqSupply -> Stream IO RawCmmGroup () -> IO UniqSupply

Test-only exports: see trac #12744

cmmNativeGen

Arguments

:: (Outputable statics, Outputable instr, Instruction instr)

=> DynFlags

-> Module

-> ModLocation

-> NcgImpl statics instr jumpDest

-> UniqSupply

-> DwarfFiles

-> LabelMap DebugBlock

-> RawCmmDecl
the cmm to generate code for

-> Int
sequence number of this top thing

-> IO (UniqSupply, DwarfFiles, [NatCmmDecl statics instr], [CLabel], Maybe [RegAllocStats statics instr], Maybe [RegAllocStats], LabelMap [UnwindPoint])

Complete native code generation phase for a single top-level chunk of Cmm. Dumping the output of each stage along the way. Global conflict graph and NGC stats

data NcgImpl statics instr jumpDest

Constructors

NcgImpl

Fields

cmmTopCodeGen :: RawCmmDecl -> NatM [NatCmmDecl statics instr]

generateJumpTableForInstr :: instr -> Maybe (NatCmmDecl statics instr)

getJumpDestBlockId :: jumpDest -> Maybe BlockId

canShortcut :: instr -> Maybe jumpDest

shortcutStatics :: (BlockId -> Maybe jumpDest) -> statics -> statics

shortcutJump :: (BlockId -> Maybe jumpDest) -> instr -> instr

pprNatCmmDecl :: NatCmmDecl statics instr -> SDoc

maxSpillSlots :: Int

allocatableRegs :: [RealReg]

ncg_x86fp_kludge :: [NatCmmDecl statics instr] -> [NatCmmDecl statics instr]

ncgExpandTop :: [NatCmmDecl statics instr] -> [NatCmmDecl statics instr]

ncgAllocMoreStack :: Int -> NatCmmDecl statics instr -> UniqSM (NatCmmDecl statics instr)

ncgMakeFarBranches :: LabelMap CmmStatics -> [NatBasicBlock instr] -> [NatBasicBlock instr]

extractUnwindPoints :: [instr] -> [UnwindPoint]

given the instruction sequence of a block, produce a list of the block's UnwindPoints See Note [What is this unwinding business?] in Debug and Note [Unwinding information in the NCG] in this module.

x86NcgImpl :: DynFlags -> NcgImpl (Alignment, CmmStatics) Instr JumpDest