Avail

Documentation

type Avails = [AvailInfo]

A collection of AvailInfo - several things that are "available"

data AvailInfo

Records what things are "available", i.e. in scope

Constructors

Avail Name
An ordinary identifier in scope

AvailTC Name [Name] [FieldLabel]
A type or class in scope. Parameters:

1) The name of the type or class 2) The available pieces of type or class, excluding field selectors. 3) The record fields of the type (see Note [Representing fields in AvailInfo]).

The AvailTC Invariant: * If the type or class is itself to be in scope, it must be *first* in this list. Thus, typically: AvailTC Eq [Eq, ==, /=]

avail :: Name -> AvailInfo

availsToNameSet :: [AvailInfo] -> NameSet

availsToNameSetWithSelectors :: [AvailInfo] -> NameSet

availsToNameEnv :: [AvailInfo] -> NameEnv AvailInfo

availName :: AvailInfo -> Name

Just the main name made available, i.e. not the available pieces of type or class brought into scope by the GenAvailInfo

availNames :: AvailInfo -> [Name]

All names made available by the availability information (excluding overloaded selectors)

availNonFldNames :: AvailInfo -> [Name]

Names for non-fields made available by the availability information

availNamesWithSelectors :: AvailInfo -> [Name]

All names made available by the availability information (including overloaded selectors)

availFlds :: AvailInfo -> [FieldLabel]

Fields made available by the availability information

stableAvailCmp :: AvailInfo -> AvailInfo -> Ordering

Compare lexicographically

plusAvail :: AvailInfo -> AvailInfo -> AvailInfo

trimAvail :: AvailInfo -> Name -> AvailInfo

trims an AvailInfo to keep only a single name

filterAvail :: (Name -> Bool) -> AvailInfo -> [AvailInfo] -> [AvailInfo]

filters an AvailInfo by the given predicate

filterAvails :: (Name -> Bool) -> [AvailInfo] -> [AvailInfo]

filters AvailInfos by the given predicate

nubAvails :: [AvailInfo] -> [AvailInfo]

Combines AvailInfos from the same family avails may have several items with the same availName E.g import Ix( Ix(..), index ) will give Ix(Ix,index,range) and Ix(index) We want to combine these; addAvail does that