BasicTypes

Documentation

type Version = Int

bumpVersion :: Version -> Version

initialVersion :: Version

type Arity = Int

type Alignment = Int

data FunctionOrData

Constructors

IsFunction

IsData

data WarningTxt

Constructors

WarningTxt [FastString]

DeprecatedTxt [FastString]

data Fixity

Constructors

Fixity Int FixityDirection

data FixityDirection

Constructors

InfixL

InfixR

InfixN

defaultFixity :: Fixity

maxPrecedence :: Int

negateFixity :: Fixity

funTyFixity :: Fixity

compareFixity :: Fixity -> Fixity -> (Bool, Bool)

newtype IPName name

Constructors

IPName name

ipNameName :: IPName name -> name

mapIPName :: (a -> b) -> IPName a -> IPName b

data RecFlag

Constructors

Recursive

NonRecursive

isRec :: RecFlag -> Bool

isNonRec :: RecFlag -> Bool

boolToRecFlag :: Bool -> RecFlag

type RuleName = FastString

data TopLevelFlag

Constructors

TopLevel

NotTopLevel

isTopLevel :: TopLevelFlag -> Bool

isNotTopLevel :: TopLevelFlag -> Bool

data OverlapFlag

Constructors

NoOverlap
This instance must not overlap another

Fields

isSafeOverlap :: Bool

OverlapOk
Silently ignore this instance if you find a more specific one that matches the constraint you are trying to resolve

Example: constraint (Foo [Int]) instances (Foo [Int]) (Foo [a]) OverlapOk Since the second instance has the OverlapOk flag, the first instance will be chosen (otherwise its ambiguous which to choose)

Fields

isSafeOverlap :: Bool

Incoherent
Like OverlapOk, but also ignore this instance if it doesn't match the constraint you are trying to resolve, but could match if the type variables in the constraint were instantiated

Example: constraint (Foo [b]) instances (Foo [Int]) Incoherent (Foo [a]) Without the Incoherent flag, we'd complain that instantiating b would change which instance was chosen

Fields

isSafeOverlap :: Bool

data Boxity

Constructors

Boxed

Unboxed

isBoxed :: Boxity -> Bool

data TupleSort

Constructors

BoxedTuple

UnboxedTuple

ConstraintTuple

tupleSortBoxity :: TupleSort -> Boxity

boxityNormalTupleSort :: Boxity -> TupleSort

tupleParens :: TupleSort -> SDoc -> SDoc

data OccInfo

Identifier occurrence information

Constructors

NoOccInfo
There are many occurrences, or unknown occurences

IAmDead
Marks unused variables. Sometimes useful for lambda and case-bound variables.

OneOcc !InsideLam !OneBranch !InterestingCxt
Occurs exactly once, not inside a rule

IAmALoopBreaker !RulesOnly
This identifier breaks a loop of mutually recursive functions. The field marks whether it is only a loop breaker due to a reference in a rule

seqOccInfo :: OccInfo -> ()

zapFragileOcc :: OccInfo -> OccInfo

isOneOcc :: OccInfo -> Bool

isDeadOcc :: OccInfo -> Bool

isStrongLoopBreaker :: OccInfo -> Bool

isWeakLoopBreaker :: OccInfo -> Bool

isNoOcc :: OccInfo -> Bool

strongLoopBreaker :: OccInfo

weakLoopBreaker :: OccInfo

type InsideLam = Bool

insideLam :: InsideLam

notInsideLam :: InsideLam

type OneBranch = Bool

oneBranch :: OneBranch

notOneBranch :: OneBranch

type InterestingCxt = Bool

data EP a

Constructors

EP

Fields

fromEP :: a

toEP :: a

data HsBang

Constructors

HsNoBang

HsStrict

HsUnpack

HsUnpackFailed

HsNoUnpack

isBanged :: HsBang -> Bool

isMarkedUnboxed :: HsBang -> Bool

data StrictnessMark

Constructors

MarkedStrict

NotMarkedStrict

isMarkedStrict :: StrictnessMark -> Bool

data DefMethSpec

Constructors

NoDM

VanillaDM

GenericDM

data CompilerPhase

Constructors

Phase PhaseNum

InitialPhase

type PhaseNum = Int

data Activation

Constructors

NeverActive

AlwaysActive

ActiveBefore PhaseNum

ActiveAfter PhaseNum

isActive :: CompilerPhase -> Activation -> Bool

isActiveIn :: PhaseNum -> Activation -> Bool

isNeverActive :: Activation -> Bool

isAlwaysActive :: Activation -> Bool

isEarlyActive :: Activation -> Bool

data RuleMatchInfo

Constructors

ConLike

FunLike

isConLike :: RuleMatchInfo -> Bool

isFunLike :: RuleMatchInfo -> Bool

data InlineSpec

Constructors

Inline

Inlinable

NoInline

EmptyInlineSpec

data InlinePragma

Constructors

InlinePragma

Fields

inl_inline :: InlineSpec

inl_sat :: Maybe Arity

inl_act :: Activation

inl_rule :: RuleMatchInfo

defaultInlinePragma :: InlinePragma

alwaysInlinePragma :: InlinePragma

neverInlinePragma :: InlinePragma

dfunInlinePragma :: InlinePragma

isDefaultInlinePragma :: InlinePragma -> Bool

isInlinePragma :: InlinePragma -> Bool

isInlinablePragma :: InlinePragma -> Bool

isAnyInlinePragma :: InlinePragma -> Bool

inlinePragmaSpec :: InlinePragma -> InlineSpec

inlinePragmaSat :: InlinePragma -> Maybe Arity

inlinePragmaActivation :: InlinePragma -> Activation

inlinePragmaRuleMatchInfo :: InlinePragma -> RuleMatchInfo

setInlinePragmaActivation :: InlinePragma -> Activation -> InlinePragma

setInlinePragmaRuleMatchInfo :: InlinePragma -> RuleMatchInfo -> InlinePragma

data SuccessFlag

Constructors

Succeeded

Failed

succeeded :: SuccessFlag -> Bool

failed :: SuccessFlag -> Bool

successIf :: Bool -> SuccessFlag

data FractionalLit

Constructors

FL

Fields

fl_text :: String

fl_value :: Rational

negateFractionalLit :: FractionalLit -> FractionalLit

integralFractionalLit :: Integer -> FractionalLit