BinIface

Description

Binary interface file support.

Documentation

writeBinIface :: DynFlags -> FilePath -> ModIface -> IO ()

Write an interface file

readBinIface :: CheckHiWay -> TraceBinIFaceReading -> FilePath -> TcRnIf a b ModIface

Read an interface file

getSymtabName :: NameCacheUpdater -> Dictionary -> SymbolTable -> BinHandle -> IO Name

getDictFastString :: Dictionary -> BinHandle -> IO FastString

data CheckHiWay

Constructors

CheckHiWay

IgnoreHiWay

data TraceBinIFaceReading

Constructors

TraceBinIFaceReading

QuietBinIFaceReading

getWithUserData :: Binary a => NameCacheUpdater -> BinHandle -> IO a

This performs a get action after reading the dictionary and symbol table. It is necessary to run this before trying to deserialise any Names or FastStrings.

putWithUserData :: Binary a => (SDoc -> IO ()) -> BinHandle -> a -> IO ()

Put a piece of data with an initialised UserData field. This is necessary if you want to serialise Names or FastStrings. It also writes a symbol table and the dictionary. This segment should be read using getWithUserData.