Binary

Documentation

data Bin a

class Binary a where

Minimal complete definition

get

Methods

put_ :: BinHandle -> a -> IO ()

put :: BinHandle -> a -> IO (Bin a)

get :: BinHandle -> IO a

data BinHandle

type SymbolTable = Array Int Name

type Dictionary = Array Int FastString

openBinMem :: Int -> IO BinHandle

seekBin :: BinHandle -> Bin a -> IO ()

seekBy :: BinHandle -> Int -> IO ()

tellBin :: BinHandle -> IO (Bin a)

castBin :: Bin a -> Bin b

isEOFBin :: BinHandle -> IO Bool

withBinBuffer :: BinHandle -> (ByteString -> IO a) -> IO a

Get access to the underlying buffer.

It is quite important that no references to the ByteString leak out of the continuation lest terrible things happen.

writeBinMem :: BinHandle -> FilePath -> IO ()

readBinMem :: FilePath -> IO BinHandle

putAt :: Binary a => BinHandle -> Bin a -> a -> IO ()

getAt :: Binary a => BinHandle -> Bin a -> IO a

For writing instances

putByte :: BinHandle -> Word8 -> IO ()

getByte :: BinHandle -> IO Word8

Lazy Binary I/O

lazyGet :: Binary a => BinHandle -> IO a

lazyPut :: Binary a => BinHandle -> a -> IO ()

User data

data UserData

Information we keep around during interface file serialization/deserialization. Namely we keep the functions for serializing and deserializing Names and FastStrings. We do this because we actually use serialization in two distinct settings,

When serializing interface files themselves

When computing the fingerprint of an IfaceDecl (which we computing by hashing its Binary serialization)

These two settings have different needs while serializing Names:

Names in interface files are serialized via a symbol table (see Note [Symbol table representation of names] in BinIface).

During fingerprinting a binding Name is serialized as the OccName and a non-binding Name is serialized as the fingerprint of the thing they represent. See Note [Fingerprinting IfaceDecls] for further discussion.

Constructors

UserData

Fields

ud_get_name :: BinHandle -> IO Name

ud_get_fs :: BinHandle -> IO FastString

ud_put_nonbinding_name :: BinHandle -> Name -> IO ()

serialize a non-binding Name (e.g. a reference to another binding).

ud_put_binding_name :: BinHandle -> Name -> IO ()

serialize a binding Name (e.g. the name of an IfaceDecl)

ud_put_fs :: BinHandle -> FastString -> IO ()

getUserData :: BinHandle -> UserData

setUserData :: BinHandle -> UserData -> BinHandle

newReadState

Arguments

:: (BinHandle -> IO Name)
how to deserialize Names

-> (BinHandle -> IO FastString)

-> UserData

newWriteState

Arguments

:: (BinHandle -> Name -> IO ())
how to serialize non-binding Names

-> (BinHandle -> Name -> IO ())
how to serialize binding Names

-> (BinHandle -> FastString -> IO ())

-> UserData

putDictionary :: BinHandle -> Int -> UniqFM (Int, FastString) -> IO ()

getDictionary :: BinHandle -> IO Dictionary

putFS :: BinHandle -> FastString -> IO ()