BuildTyCl

Documentation

buildDataCon :: FamInstEnvs -> Name -> Bool -> TyConRepName -> [HsSrcBang] -> Maybe [HsImplBang] -> [FieldLabel] -> [TyVarBinder] -> [TyVarBinder] -> [EqSpec] -> ThetaType -> [Type] -> Type -> TyCon -> TcRnIf m n DataCon

mkDataConUnivTyVarBinders :: [TyConBinder] -> [TyVarBinder]

buildPatSyn

Arguments

:: Name

-> Bool

-> (Id, Bool)

-> Maybe (Id, Bool)

-> ([TyVarBinder], ThetaType)
Univ and req

-> ([TyVarBinder], ThetaType)
Ex and prov

-> [Type]
Argument types

-> Type
Result type

-> [FieldLabel]
Field labels for a record pattern synonym

-> PatSyn

type TcMethInfo = (Name, Type, Maybe (DefMethSpec (SrcSpan, Type)))

buildClass :: Name -> [TyConBinder] -> [Role] -> [FunDep TyVar] -> Maybe (ThetaType, [ClassATItem], [TcMethInfo], ClassMinimalDef) -> TcRnIf m n Class

mkNewTyConRhs :: Name -> TyCon -> DataCon -> TcRnIf m n AlgTyConRhs

Monadic because it makes a Name for the coercion TyCon We pass the Name of the parent TyCon, as well as the TyCon itself, because the latter is part of a knot, whereas the former is not.

mkDataTyConRhs :: [DataCon] -> AlgTyConRhs

newImplicitBinder :: Name -> (OccName -> OccName) -> TcRnIf m n Name

newTyConRepName :: Name -> TcRnIf gbl lcl TyConRepName

Make the TyConRepName for this TyCon