Cmm

Cmm top-level datatypes

type CmmProgram = [CmmGroup]

type CmmGroup = GenCmmGroup CmmStatics CmmTopInfo CmmGraph

type GenCmmGroup d h g = [GenCmmDecl d h g]

type CmmDecl = GenCmmDecl CmmStatics CmmTopInfo CmmGraph

data GenCmmDecl d h g

A top-level chunk, abstracted over the type of the contents of the basic blocks (Cmm or instructions are the likely instantiations).

Constructors

CmmProc h CLabel [GlobalReg] g

CmmData Section d

type CmmGraph = GenCmmGraph CmmNode

data GenCmmGraph n

Constructors

CmmGraph

Fields

g_entry :: BlockId

g_graph :: Graph n C C

type CmmBlock = Block CmmNode C C

type RawCmmDecl = GenCmmDecl CmmStatics (BlockEnv CmmStatics) CmmGraph

type RawCmmGroup = GenCmmGroup CmmStatics (BlockEnv CmmStatics) CmmGraph

data Section

Constructors

Text

Data

ReadOnlyData

RelocatableReadOnlyData

UninitialisedData

ReadOnlyData16

OtherSection String

data CmmStatics

Constructors

Statics CLabel [CmmStatic]

data CmmStatic

Constructors

CmmStaticLit CmmLit

CmmUninitialised Int

CmmString [Word8]

Blocks containing lists

data GenBasicBlock i

Constructors

BasicBlock BlockId [i]

blockId :: GenBasicBlock i -> BlockId

The branch block id is that of the first block in the branch, which is that branch's entry point

newtype ListGraph i

Constructors

ListGraph [GenBasicBlock i]

pprBBlock :: Outputable stmt => GenBasicBlock stmt -> SDoc

Cmm graphs

type CmmReplGraph e x = GenCmmReplGraph CmmNode e x

type GenCmmReplGraph n e x = UniqSM (Maybe (Graph n e x))

type CmmFwdRewrite f = FwdRewrite UniqSM CmmNode f

type CmmBwdRewrite f = BwdRewrite UniqSM CmmNode f

Info Tables

data CmmTopInfo

Constructors

TopInfo

Fields

info_tbls :: BlockEnv CmmInfoTable

stack_info :: CmmStackInfo

data CmmStackInfo

Constructors

StackInfo

Fields

arg_space :: ByteOff

updfr_space :: Maybe ByteOff

do_layout :: Bool

data CmmInfoTable

Info table as a haskell data type

Constructors

CmmInfoTable

Fields

cit_lbl :: CLabel

cit_rep :: SMRep

cit_prof :: ProfilingInfo

cit_srt :: C_SRT

topInfoTable :: GenCmmDecl a CmmTopInfo (GenCmmGraph n) -> Maybe CmmInfoTable

data ClosureTypeInfo

Constructors

Constr ConstrTag ConstrDescription

Fun FunArity ArgDescr

Thunk

ThunkSelector SelectorOffset

BlackHole

IndStatic

data C_SRT

Constructors

NoC_SRT

C_SRT !CLabel !WordOff !StgHalfWord

needsSRT :: C_SRT -> Bool

data ProfilingInfo

Constructors

NoProfilingInfo

ProfilingInfo [Word8] [Word8]

type ConstrDescription = [Word8]

Statements, expressions and types

module CmmNode

module CmmExpr