CmmCallConv

Documentation

data ParamLocation

Constructors

RegisterParam GlobalReg

StackParam ByteOff

assignArgumentsPos :: Convention -> (a -> CmmType) -> [a] -> [(a, ParamLocation)]

JD: For the new stack story, I want arguments passed on the stack to manifest as positive offsets in a CallArea, not negative offsets from the stack pointer. Also, I want byte offsets, not word offsets.