CmmExpr

Documentation

data CmmExpr

Constructors

CmmLit CmmLit

CmmLoad !CmmExpr !CmmType

CmmReg !CmmReg

CmmMachOp MachOp [CmmExpr]

CmmStackSlot Area !Int

CmmRegOff !CmmReg Int

cmmExprType :: DynFlags -> CmmExpr -> CmmType

cmmExprWidth :: DynFlags -> CmmExpr -> Width

maybeInvertCmmExpr :: CmmExpr -> Maybe CmmExpr

data CmmReg

Constructors

CmmLocal !LocalReg

CmmGlobal GlobalReg

cmmRegType :: DynFlags -> CmmReg -> CmmType

data CmmLit

Constructors

CmmInt !Integer Width

CmmFloat Rational Width

CmmVec [CmmLit]

CmmLabel CLabel

CmmLabelOff CLabel Int

CmmLabelDiffOff CLabel CLabel Int

CmmBlock !BlockId

CmmHighStackMark

cmmLitType :: DynFlags -> CmmLit -> CmmType

data LocalReg

Constructors

LocalReg !Unique CmmType
Parameters: 1. Identifier 2. Type

localRegType :: LocalReg -> CmmType

data GlobalReg

Constructors

VanillaReg !Int VGcPtr

FloatReg !Int

DoubleReg !Int

LongReg !Int

XmmReg !Int

YmmReg !Int

ZmmReg !Int

Sp

SpLim

Hp

HpLim

CCCS

CurrentTSO

CurrentNursery

HpAlloc

EagerBlackholeInfo

GCEnter1

GCFun

BaseReg

MachSp

UnwindReturnReg

PicBaseReg

isArgReg :: GlobalReg -> Bool

globalRegType :: DynFlags -> GlobalReg -> CmmType

spReg :: CmmReg

hpReg :: CmmReg

spLimReg :: CmmReg

nodeReg :: CmmReg

node :: GlobalReg

baseReg :: CmmReg

data VGcPtr

Constructors

VGcPtr

VNonGcPtr

class Ord r => DefinerOfRegs r a where

Minimal complete definition

foldRegsDefd

Methods

foldRegsDefd :: DynFlags -> (b -> r -> b) -> b -> a -> b

class Ord r => UserOfRegs r a where

Minimal complete definition

foldRegsUsed

Methods

foldRegsUsed :: DynFlags -> (b -> r -> b) -> b -> a -> b

foldRegsDefd :: DefinerOfRegs r a => DynFlags -> (b -> r -> b) -> b -> a -> b

foldRegsUsed :: UserOfRegs r a => DynFlags -> (b -> r -> b) -> b -> a -> b

foldLocalRegsDefd :: DefinerOfRegs LocalReg a => DynFlags -> (b -> LocalReg -> b) -> b -> a -> b

foldLocalRegsUsed :: UserOfRegs LocalReg a => DynFlags -> (b -> LocalReg -> b) -> b -> a -> b

type RegSet r = Set r

Sets of registers

type LocalRegSet = RegSet LocalReg

type GlobalRegSet = RegSet GlobalReg

emptyRegSet :: RegSet r

elemRegSet :: Ord r => r -> RegSet r -> Bool

extendRegSet :: Ord r => RegSet r -> r -> RegSet r

deleteFromRegSet :: Ord r => RegSet r -> r -> RegSet r

mkRegSet :: Ord r => [r] -> RegSet r

plusRegSet :: Ord r => RegSet r -> RegSet r -> RegSet r

minusRegSet :: Ord r => RegSet r -> RegSet r -> RegSet r

timesRegSet :: Ord r => RegSet r -> RegSet r -> RegSet r

sizeRegSet :: RegSet r -> Int

nullRegSet :: RegSet r -> Bool

regSetToList :: RegSet r -> [r]

data Area

A stack area is either the stack slot where a variable is spilled or the stack space where function arguments and results are passed.

Constructors

Old

Young !BlockId

module CmmMachOp

module CmmType