CmmExpr

Documentation

data CmmExpr

Constructors

CmmLit CmmLit

CmmLoad CmmExpr CmmType

CmmReg CmmReg

CmmMachOp MachOp [CmmExpr]

CmmStackSlot Area Int

CmmRegOff CmmReg Int

cmmExprType :: CmmExpr -> CmmType

cmmExprWidth :: CmmExpr -> Width

maybeInvertCmmExpr :: CmmExpr -> Maybe CmmExpr

data CmmReg

Constructors

CmmLocal LocalReg

CmmGlobal GlobalReg

cmmRegType :: CmmReg -> CmmType

data CmmLit

Constructors

CmmInt Integer Width

CmmFloat Rational Width

CmmLabel CLabel

CmmLabelOff CLabel Int

CmmLabelDiffOff CLabel CLabel Int

CmmBlock BlockId

CmmHighStackMark

cmmLitType :: CmmLit -> CmmType

data LocalReg

Constructors

LocalReg !Unique CmmType
Parameters: 1. Identifier 2. Type

localRegType :: LocalReg -> CmmType

data GlobalReg

Constructors

VanillaReg !Int VGcPtr

FloatReg !Int

DoubleReg !Int

LongReg !Int

Sp

SpLim

Hp

HpLim

CCCS

CurrentTSO

CurrentNursery

HpAlloc

EagerBlackholeInfo

GCEnter1

GCFun

BaseReg

PicBaseReg

globalRegType :: GlobalReg -> CmmType

spReg :: CmmReg

hpReg :: CmmReg

spLimReg :: CmmReg

nodeReg :: CmmReg

node :: GlobalReg

baseReg :: CmmReg

data VGcPtr

Constructors

VGcPtr

VNonGcPtr

vgcFlag :: CmmType -> VGcPtr

class DefinerOfLocalRegs a where

Methods

foldRegsDefd :: (b -> LocalReg -> b) -> b -> a -> b

class UserOfLocalRegs a where

Methods

foldRegsUsed :: (b -> LocalReg -> b) -> b -> a -> b

filterRegsUsed :: UserOfLocalRegs e => (LocalReg -> Bool) -> e -> RegSet

class DefinerOfSlots a where

Methods

foldSlotsDefd :: (b -> SubArea -> b) -> b -> a -> b

class UserOfSlots a where

Methods

foldSlotsUsed :: (b -> SubArea -> b) -> b -> a -> b

type RegSet = UniqSet LocalReg

Sets of local registers

emptyRegSet :: RegSet

elemRegSet :: LocalReg -> RegSet -> Bool

extendRegSet :: RegSet -> LocalReg -> RegSet

deleteFromRegSet :: RegSet -> LocalReg -> RegSet

mkRegSet :: [LocalReg] -> RegSet

plusRegSet :: RegSet -> RegSet -> RegSet

minusRegSet :: RegSet -> RegSet -> RegSet

timesRegSet :: RegSet -> RegSet -> RegSet

regUsedIn :: CmmReg -> CmmExpr -> Bool

regSlot :: LocalReg -> CmmExpr

data Area

A stack area is either the stack slot where a variable is spilled or the stack space where function arguments and results are passed.

Constructors

RegSlot LocalReg

CallArea AreaId

data AreaId

Constructors

Old

Young BlockId

type SubArea = (Area, Int, Int)

type SubAreaSet = Map Area [SubArea]

type AreaMap = Map Area Int

isStackSlotOf :: CmmExpr -> LocalReg -> Bool

module CmmMachOp

module CmmType