CoreSyn

Description

CoreSyn holds all the main data types for use by for the Glasgow Haskell Compiler midsection

Main data types

data Expr b

This is the data type that represents GHCs core intermediate language. Currently GHC uses System FC http://research.microsoft.com/~simonpj/papers/ext-f/ for this purpose, which is closely related to the simpler and better known System F http://en.wikipedia.org/wiki/System_F.

We get from Haskell source to this Core language in a number of stages:

The source code is parsed into an abstract syntax tree, which is represented by the data type HsExpr with the names being RdrNames

This syntax tree is renamed, which attaches a Unique to every RdrName (yielding a Name) to disambiguate identifiers which are lexically identical. For example, this program:

      f x = let f x = x + 1
            in f (x - 2)

Would be renamed by having Uniques attached so it looked something like this:

      f_1 x_2 = let f_3 x_4 = x_4 + 1
                in f_3 (x_2 - 2)

The resulting syntax tree undergoes type checking (which also deals with instantiating type class arguments) to yield a HsExpr type that has Id as it's names.

Finally the syntax tree is desugared from the expressive HsExpr type into this Expr type, which has far fewer constructors and hence is easier to perform optimization, analysis and code generation on.

The type parameter b is for the type of binders in the expression tree.

The language consists of the following elements:

Variables

Primitive literals

Applications: note that the argument may be a Type.

See CoreSyn for another invariant

Lambda abstraction

Recursive and non recursive lets. Operationally this corresponds to allocating a thunk for the things bound and then executing the sub-expression.

The right hand sides of all top-level and recursive lets must be of lifted type (see Type for the meaning of lifted vs. unlifted).

The right hand side of of a non-recursive Let _and_ the argument of an App, may be of unlifted type, but only if the expression is ok-for-speculation. This means that the let can be floated around without difficulty. For example, this is OK:

 y::Int# = x +# 1#

But this is not, as it may affect termination if the expression is floated out:

 y::Int# = fac 4#

In this situation you should use case rather than a let. The function needsCaseBinding can help you determine which to generate, or alternatively use mkCoreLet rather than this constructor directly, which will generate a case if necessary

We allow a non-recursive let to bind a type variable, thus:

 Let (NonRec tv (Type ty)) body

This can be very convenient for postponing type substitutions until the next run of the simplifier.

At the moment, the rest of the compiler only deals with type-let in a Let expression, rather than at top level. We may want to revist this choice.

Case split. Operationally this corresponds to evaluating the scrutinee (expression examined) to weak head normal form and then examining at most one level of resulting constructor (i.e. you cannot do nested pattern matching directly with this).

The binder gets bound to the value of the scrutinee, and the Type must be that of all the case alternatives

This is one of the more complicated elements of the Core language, and comes with a number of restrictions:

The list of alternatives is non-empty

The DEFAULT case alternative must be first in the list, if it occurs at all.

The remaining cases are in order of increasing tag (for DataAlts) or lit (for LitAlts). This makes finding the relevant constructor easy, and makes comparison easier too.

The list of alternatives must be exhaustive. An exhaustive case does not necessarily mention all constructors:

    	      data Foo = Red | Green | Blue
    	 ... case x of 
    	      Red   -> True
    	      other -> f (case x of 
    	                      Green -> ...
    	                      Blue  -> ... ) ...

The inner case does not need a Red alternative, because x can't be Red at that program point.

Cast an expression to a particular type. This is used to implement newtypes (a newtype constructor or destructor just becomes a Cast in Core) and GADTs.

Notes. These allow general information to be added to expressions in the syntax tree

A type: this should only show up at the top level of an Arg

A coercion

Constructors

Var Id

Lit Literal

App (Expr b) (Arg b)

Lam b (Expr b)

Let (Bind b) (Expr b)

Case (Expr b) b Type [Alt b]

Cast (Expr b) Coercion

Tick (Tickish Id) (Expr b)

Type Type

Coercion Coercion

type Alt b = (AltCon, [b], Expr b)

A case split alternative. Consists of the constructor leading to the alternative, the variables bound from the constructor, and the expression to be executed given that binding. The default alternative is (DEFAULT, [], rhs)

data Bind b

Binding, used for top level bindings in a module and local bindings in a let.

Constructors

NonRec b (Expr b)

Rec [(b, Expr b)]

data AltCon

A case alternative constructor (i.e. pattern match)

Constructors

DataAlt DataCon

LitAlt Literal
A literal: case e of { 1 -> ... } Invariant: always an *unlifted* literal See Note [Literal alternatives]

DEFAULT
Trivial alternative: case e of { _ -> ... }

type Arg b = Expr b

Type synonym for expressions that occur in function argument positions. Only Arg should contain a Type at top level, general Expr should not

data Tickish id

Allows attaching extra information to points in expressions

Constructors

ProfNote
An {--} profiling annotation, either automatically added by the desugarer as a result of -auto-all, or added by the user.

Fields

profNoteCC :: CostCentre

the cost centre

profNoteCount :: !Bool

bump the entry count?

profNoteScope :: !Bool

scopes over the enclosed expression (i.e. not just a tick)

HpcTick
A tick used by HPC to track the execution of each subexpression in the original source code.

Fields

tickModule :: Module

tickId :: !Int

Breakpoint
A breakpoint for the GHCi debugger. This behaves like an HPC tick, but has a list of free variables which will be available for inspection in GHCi when the program stops at the breakpoint.

NB. we must take account of these Ids when (a) counting free variables, and (b) substituting (don't substitute for them)

Fields

breakpointId :: !Int

breakpointFVs :: [id]

the order of this list is important: it matches the order of the lists in the appropriate entry in HscTypes.ModBreaks.

Careful about substitution! See Note [substTickish] in CoreSubst.

type CoreProgram = [CoreBind]

type CoreExpr = Expr CoreBndr

Expressions where binders are CoreBndrs

type CoreAlt = Alt CoreBndr

Case alternatives where binders are CoreBndrs

type CoreBind = Bind CoreBndr

Binding groups where binders are CoreBndrs

type CoreArg = Arg CoreBndr

Argument expressions where binders are CoreBndrs

type CoreBndr = Var

The common case for the type of binders and variables when we are manipulating the Core language within GHC

type TaggedExpr t = Expr (TaggedBndr t)

type TaggedAlt t = Alt (TaggedBndr t)

type TaggedBind t = Bind (TaggedBndr t)

type TaggedArg t = Arg (TaggedBndr t)

data TaggedBndr t

Binders are tagged with a t

Constructors

TB CoreBndr t

Expr construction

mkLets :: [Bind b] -> Expr b -> Expr b

Bind all supplied binding groups over an expression in a nested let expression. Prefer to use mkCoreLets if possible

mkLams :: [b] -> Expr b -> Expr b

Bind all supplied binders over an expression in a nested lambda expression. Prefer to use mkCoreLams if possible

mkApps :: Expr b -> [Arg b] -> Expr b

Apply a list of argument expressions to a function expression in a nested fashion. Prefer to use mkCoreApps if possible

mkTyApps :: Expr b -> [Type] -> Expr b

Apply a list of type argument expressions to a function expression in a nested fashion

mkCoApps :: Expr b -> [Coercion] -> Expr b

Apply a list of coercion argument expressions to a function expression in a nested fashion

mkVarApps :: Expr b -> [Var] -> Expr b

Apply a list of type or value variables to a function expression in a nested fashion

mkIntLit :: Integer -> Expr b

Create a machine integer literal expression of type Int# from an Integer. If you want an expression of type Int use mkIntExpr

mkIntLitInt :: Int -> Expr b

Create a machine integer literal expression of type Int# from an Int. If you want an expression of type Int use mkIntExpr

mkWordLit :: Integer -> Expr b

Create a machine word literal expression of type Word# from an Integer. If you want an expression of type Word use mkWordExpr

mkWordLitWord :: Word -> Expr b

Create a machine word literal expression of type Word# from a Word. If you want an expression of type Word use mkWordExpr

mkWord64LitWord64 :: Word64 -> Expr b

mkInt64LitInt64 :: Int64 -> Expr b

mkCharLit :: Char -> Expr b

Create a machine character literal expression of type Char#. If you want an expression of type Char use mkCharExpr

mkStringLit :: String -> Expr b

Create a machine string literal expression of type Addr#. If you want an expression of type String use mkStringExpr

mkFloatLit :: Rational -> Expr b

Create a machine single precision literal expression of type Float# from a Rational. If you want an expression of type Float use mkFloatExpr

mkFloatLitFloat :: Float -> Expr b

Create a machine single precision literal expression of type Float# from a Float. If you want an expression of type Float use mkFloatExpr

mkDoubleLit :: Rational -> Expr b

Create a machine double precision literal expression of type Double# from a Rational. If you want an expression of type Double use mkDoubleExpr

mkDoubleLitDouble :: Double -> Expr b

Create a machine double precision literal expression of type Double# from a Double. If you want an expression of type Double use mkDoubleExpr

mkConApp :: DataCon -> [Arg b] -> Expr b

Apply a list of argument expressions to a data constructor in a nested fashion. Prefer to use mkCoreConApps if possible

mkTyBind :: TyVar -> Type -> CoreBind

Create a binding group where a type variable is bound to a type. Per CoreSyn, this can only be used to bind something in a non-recursive let expression

mkCoBind :: CoVar -> Coercion -> CoreBind

Create a binding group where a type variable is bound to a type. Per CoreSyn, this can only be used to bind something in a non-recursive let expression

varToCoreExpr :: CoreBndr -> Expr b

Convert a binder into either a Var or Type Expr appropriately

varsToCoreExprs :: [CoreBndr] -> [Expr b]

isId :: Var -> Bool

cmpAltCon :: AltCon -> AltCon -> Ordering

Compares AltCons within a single list of alternatives

cmpAlt :: Alt b -> Alt b -> Ordering

ltAlt :: Alt b -> Alt b -> Bool

Simple Expr access functions and predicates

bindersOf :: Bind b -> [b]

Extract every variable by this group

bindersOfBinds :: [Bind b] -> [b]

bindersOf applied to a list of binding groups

rhssOfBind :: Bind b -> [Expr b]

rhssOfAlts :: [Alt b] -> [Expr b]

collectBinders :: Expr b -> ([b], Expr b)

We often want to strip off leading lambdas before getting down to business. This function is your friend.

collectTyBinders :: CoreExpr -> ([TyVar], CoreExpr)

Collect as many type bindings as possible from the front of a nested lambda

collectValBinders :: CoreExpr -> ([Id], CoreExpr)

Collect as many value bindings as possible from the front of a nested lambda

collectTyAndValBinders :: CoreExpr -> ([TyVar], [Id], CoreExpr)

Collect type binders from the front of the lambda first, then follow up by collecting as many value bindings as possible from the resulting stripped expression

collectArgs :: Expr b -> (Expr b, [Arg b])

Takes a nested application expression and returns the the function being applied and the arguments to which it is applied

flattenBinds :: [Bind b] -> [(b, Expr b)]

Collapse all the bindings in the supplied groups into a single list of lhs/rhs pairs suitable for binding in a Rec binding group

isValArg :: Expr b -> Bool

Returns False iff the expression is a Type or Coercion expression at its top level

isTypeArg :: Expr b -> Bool

Returns True iff the expression is a Type expression at its top level. Note this does NOT include Coercions.

isTyCoArg :: Expr b -> Bool

Returns True iff the expression is a Type or Coercion expression at its top level

valArgCount :: [Arg b] -> Int

The number of argument expressions that are values rather than types at their top level

valBndrCount :: [CoreBndr] -> Int

The number of binders that bind values rather than types

isRuntimeArg :: CoreExpr -> Bool

Will this argument expression exist at runtime?

isRuntimeVar :: Var -> Bool

Will this variable exist at runtime?

tickishCounts :: Tickish id -> Bool

A tick note is one that counts evaluations in some way. We cannot discard a tick, and the compiler should preserve the number of ticks as far as possible.

Hwever, we stil allow the simplifier to increase or decrease sharing, so in practice the actual number of ticks may vary, except that we never change the value from zero to non-zero or vice versa.

tickishScoped :: Tickish id -> Bool

tickishIsCode :: Tickish id -> Bool

Return True if this source annotation compiles to some code, or will disappear before the backend.

mkNoTick :: Tickish id -> Tickish id

mkNoScope :: Tickish id -> Tickish id

tickishCanSplit :: Tickish Id -> Bool

Return True if this Tick can be split into (tick,scope) parts with mkNoScope and mkNoTick respectively.

Unfolding data types

data Unfolding

Records the unfolding of an identifier, which is approximately the form the identifier would have if we substituted its definition in for the identifier. This type should be treated as abstract everywhere except in CoreUnfold

Constructors

NoUnfolding
We have no information about the unfolding

OtherCon [AltCon]
It ain't one of these constructors. OtherCon xs also indicates that something has been evaluated and hence there's no point in re-evaluating it. OtherCon [] is used even for non-data-type values to indicated evaluated-ness. Notably:

 data C = C !(Int -> Int)
 case x of { C f -> ... }

Here, f gets an OtherCon [] unfolding.

DFunUnfolding Arity DataCon [CoreExpr]

CoreUnfolding
An unfolding with redundant cached information. Parameters:

uf_tmpl: Template used to perform unfolding; NB: Occurrence info is guaranteed correct: see Note [OccInfo in unfoldings and rules]

uf_is_top: Is this a top level binding?

uf_is_value: exprIsHNF template (cached); it is ok to discard a seq on this variable

uf_is_cheap: Does this waste only a little work if we expand it inside an inlining? Basically this is a cached version of exprIsCheap

uf_guidance: Tells us about the size of the unfolding template

Fields

uf_tmpl :: CoreExpr

uf_src :: UnfoldingSource

uf_is_top :: Bool

uf_arity :: Arity

uf_is_value :: Bool

uf_is_conlike :: Bool

uf_is_cheap :: Bool

uf_expandable :: Bool

uf_guidance :: UnfoldingGuidance

data UnfoldingGuidance

UnfoldingGuidance says when unfolding should take place

Constructors

UnfWhen

Fields

ug_unsat_ok :: Bool

ug_boring_ok :: Bool

UnfIfGoodArgs

Fields

ug_args :: [Int]

ug_size :: Int

ug_res :: Int

UnfNever

data UnfoldingSource

Constructors

InlineRhs

InlineStable

InlineCompulsory

InlineWrapper Id

Constructing Unfoldings

noUnfolding :: Unfolding

There is no known Unfolding

evaldUnfolding :: Unfolding

This unfolding marks the associated thing as being evaluated

mkOtherCon :: [AltCon] -> Unfolding

unSaturatedOk :: Bool

needSaturated :: Bool

boringCxtOk :: Bool

boringCxtNotOk :: Bool

Predicates and deconstruction on Unfolding

unfoldingTemplate :: Unfolding -> CoreExpr

Retrieves the template of an unfolding: panics if none is known

setUnfoldingTemplate :: Unfolding -> CoreExpr -> Unfolding

expandUnfolding_maybe :: Unfolding -> Maybe CoreExpr

maybeUnfoldingTemplate :: Unfolding -> Maybe CoreExpr

Retrieves the template of an unfolding if possible

otherCons :: Unfolding -> [AltCon]

The constructors that the unfolding could never be: returns [] if no information is available

unfoldingArity :: Unfolding -> Arity

isValueUnfolding :: Unfolding -> Bool

Determines if it is certainly the case that the unfolding will yield a value (something in HNF): returns False if unsure

isEvaldUnfolding :: Unfolding -> Bool

Determines if it possibly the case that the unfolding will yield a value. Unlike isValueUnfolding it returns True for OtherCon

isCheapUnfolding :: Unfolding -> Bool

Is the thing we will unfold into certainly cheap?

isExpandableUnfolding :: Unfolding -> Bool

isConLikeUnfolding :: Unfolding -> Bool

True if the unfolding is a constructor application, the application of a CONLIKE function or OtherCon

isCompulsoryUnfolding :: Unfolding -> Bool

isStableUnfolding :: Unfolding -> Bool

isStableCoreUnfolding_maybe :: Unfolding -> Maybe UnfoldingSource

isClosedUnfolding :: Unfolding -> Bool

hasSomeUnfolding :: Unfolding -> Bool

Only returns False if there is no unfolding information available at all

canUnfold :: Unfolding -> Bool

neverUnfoldGuidance :: UnfoldingGuidance -> Bool

isStableSource :: UnfoldingSource -> Bool

Strictness

seqExpr :: CoreExpr -> ()

seqExprs :: [CoreExpr] -> ()

seqUnfolding :: Unfolding -> ()

Annotated expression data types

type AnnExpr bndr annot = (annot, AnnExpr' bndr annot)

Annotated core: allows annotation at every node in the tree

data AnnExpr' bndr annot

A clone of the Expr type but allowing annotation at every tree node

Constructors

AnnVar Id

AnnLit Literal

AnnLam bndr (AnnExpr bndr annot)

AnnApp (AnnExpr bndr annot) (AnnExpr bndr annot)

AnnCase (AnnExpr bndr annot) bndr Type [AnnAlt bndr annot]

AnnLet (AnnBind bndr annot) (AnnExpr bndr annot)

AnnCast (AnnExpr bndr annot) (annot, Coercion)

AnnTick (Tickish Id) (AnnExpr bndr annot)

AnnType Type

AnnCoercion Coercion

data AnnBind bndr annot

A clone of the Bind type but allowing annotation at every tree node

Constructors

AnnNonRec bndr (AnnExpr bndr annot)

AnnRec [(bndr, AnnExpr bndr annot)]

type AnnAlt bndr annot = (AltCon, [bndr], AnnExpr bndr annot)

A clone of the Alt type but allowing annotation at every tree node

Operations on annotated expressions

collectAnnArgs :: AnnExpr b a -> (AnnExpr b a, [AnnExpr b a])

Takes a nested application expression and returns the the function being applied and the arguments to which it is applied

Operations on annotations

deAnnotate :: AnnExpr bndr annot -> Expr bndr

deAnnotate' :: AnnExpr' bndr annot -> Expr bndr

deAnnAlt :: AnnAlt bndr annot -> Alt bndr

collectAnnBndrs :: AnnExpr bndr annot -> ([bndr], AnnExpr bndr annot)

As collectBinders but for AnnExpr rather than Expr

Core rule data types

data CoreRule

A CoreRule is:

"Local" if the function it is a rule for is defined in the same module as the rule itself.

"Orphan" if nothing on the LHS is defined in the same module as the rule itself

Constructors

Rule

Fields

ru_name :: RuleName

As above

Name of the rule, for communication with the user

ru_act :: Activation

When the rule is active

ru_fn :: Name

As above

Name of the Id at the head of this rule

ru_rough :: [Maybe Name]

Name at the head of each argument to the left hand side

ru_bndrs :: [CoreBndr]

Variables quantified over

ru_args :: [CoreExpr]

Left hand side arguments

ru_rhs :: CoreExpr

Right hand side of the rule Occurrence info is guaranteed correct See Note [OccInfo in unfoldings and rules]

ru_auto :: Bool

True = this rule is auto-generated False = generated at the users behest Main effect: reporting of orphan-hood

ru_local :: Bool

True iff the fn at the head of the rule is defined in the same module as the rule and is not an implicit Id (like a record selector, class operation, or data constructor)

BuiltinRule
Built-in rules are used for constant folding and suchlike. They have no free variables.

Fields

ru_name :: RuleName

As above

Name of the rule, for communication with the user

ru_fn :: Name

As above

Name of the Id at the head of this rule

ru_nargs :: Int

Number of arguments that ru_try consumes, if it fires, including type arguments

ru_try :: IdUnfoldingFun -> [CoreExpr] -> Maybe CoreExpr

This function does the rewrite. It given too many arguments, it simply discards them; the returned CoreExpr is just the rewrite of ru_fn applied to the first ru_nargs args

type RuleName = FastString

type IdUnfoldingFun = Id -> Unfolding

Operations on CoreRules

seqRules :: [CoreRule] -> ()

ruleArity :: CoreRule -> Int

The number of arguments the ru_fn must be applied to before the rule can match on it

ruleName :: CoreRule -> RuleName

ruleIdName :: CoreRule -> Name

The Name of the Id at the head of the rule left hand side

ruleActivation :: CoreRule -> Activation

setRuleIdName :: Name -> CoreRule -> CoreRule

Set the Name of the Id at the head of the rule left hand side

isBuiltinRule :: CoreRule -> Bool

isLocalRule :: CoreRule -> Bool

Core vectorisation declarations data type

data CoreVect

Constructors

Vect Id (Maybe CoreExpr)

NoVect Id

VectType Bool TyCon (Maybe TyCon)

VectClass TyCon

VectInst Id