DriverPhases

Documentation

data HscSource

Constructors

HsSrcFile

HsBootFile

HsigFile

isHsBootOrSig :: HscSource -> Bool

hscSourceString :: HscSource -> String

data Phase

Constructors

Unlit HscSource

Cpp HscSource

HsPp HscSource

Hsc HscSource

Ccxx

Cc

Cobjc

Cobjcxx

HCc

Splitter

SplitAs

As Bool

LlvmOpt

LlvmLlc

LlvmMangle

CmmCpp

Cmm

MergeForeign

StopLn

happensBefore :: DynFlags -> Phase -> Phase -> Bool

eqPhase :: Phase -> Phase -> Bool

anyHsc :: Phase

isStopLn :: Phase -> Bool

startPhase :: String -> Phase

phaseInputExt :: Phase -> String

isHaskellishSuffix :: String -> Bool

isHaskellSrcSuffix :: String -> Bool

isBackpackishSuffix :: String -> Bool

isObjectSuffix :: Platform -> String -> Bool

isCishSuffix :: String -> Bool

isDynLibSuffix :: Platform -> String -> Bool

isHaskellUserSrcSuffix :: String -> Bool

isHaskellSigSuffix :: String -> Bool

isSourceSuffix :: String -> Bool

isHaskellishTarget :: (String, Maybe Phase) -> Bool

When we are given files (modified by -x arguments) we need to determine if they are Haskellish or not to figure out how we should try to compile it. The rules are:

If no -x flag was specified, we check to see if the file looks like a module name, has no extension, or has a Haskell source extension.

If an -x flag was specified, we just make sure the specified suffix is a Haskell one.

isHaskellishFilename :: FilePath -> Bool

isHaskellSrcFilename :: FilePath -> Bool

isHaskellSigFilename :: FilePath -> Bool

isObjectFilename :: Platform -> FilePath -> Bool

isCishFilename :: FilePath -> Bool

isDynLibFilename :: Platform -> FilePath -> Bool

isHaskellUserSrcFilename :: FilePath -> Bool

isSourceFilename :: FilePath -> Bool