DsExpr

Documentation

dsExpr :: HsExpr GhcTc -> DsM CoreExpr

dsLExpr :: LHsExpr GhcTc -> DsM CoreExpr

dsLExprNoLP :: LHsExpr GhcTc -> DsM CoreExpr

Variant of dsLExpr that ensures that the result is not levity polymorphic. This should be used when the resulting expression will be an argument to some other function. See Note [Levity polymorphism checking] in DsMonad See Note [Levity polymorphism invariants] in CoreSyn

dsLocalBinds :: LHsLocalBinds GhcTc -> CoreExpr -> DsM CoreExpr

dsValBinds :: HsValBinds GhcTc -> CoreExpr -> DsM CoreExpr

dsLit :: HsLit GhcRn -> DsM CoreExpr

dsSyntaxExpr :: SyntaxExpr GhcTc -> [CoreExpr] -> DsM CoreExpr