DsExpr

Documentation

dsExpr :: HsExpr Id -> DsM CoreExpr

dsLExpr :: LHsExpr Id -> DsM CoreExpr

dsLExprNoLP :: LHsExpr Id -> DsM CoreExpr

Variant of dsLExpr that ensures that the result is not levity polymorphic. This should be used when the resulting expression will be an argument to some other function. See Note [Levity polymorphism checking] in DsMonad See Note [Levity polymorphism invariants] in CoreSyn

dsLocalBinds :: LHsLocalBinds Id -> CoreExpr -> DsM CoreExpr

dsValBinds :: HsValBinds Id -> CoreExpr -> DsM CoreExpr

dsLit :: HsLit -> DsM CoreExpr

dsSyntaxExpr :: SyntaxExpr Id -> [CoreExpr] -> DsM CoreExpr