DsUsage

Dependency/fingerprinting code (used by MkIface)

mkUsageInfo :: HscEnv -> Module -> ImportedMods -> NameSet -> [FilePath] -> [(Module, Fingerprint)] -> IO [Usage]

mkUsedNames :: TcGblEnv -> NameSet

mkDependencies :: TcGblEnv -> IO Dependencies

Extract information from the rename and typecheck phases to produce a dependencies information for the module being compiled.