FileCleanup

Documentation

data TempFileLifetime

Used when a temp file is created. This determines which component Set of FilesToClean will get the temp file

Constructors

TFL_CurrentModule
A file with lifetime TFL_CurrentModule will be cleaned up at the end of upweep_mod

TFL_GhcSession
A file with lifetime TFL_GhcSession will be cleaned up at the end of runGhc(T)

cleanTempDirs :: DynFlags -> IO ()

cleanTempFiles :: DynFlags -> IO ()

Delete all files in filesToClean dflags.

cleanCurrentModuleTempFiles :: DynFlags -> IO ()

Delete all files in filesToClean dflags. That have lifetime TFL_CurrentModule. If a file must be cleaned eventually, but must survive a cleanCurrentModuleTempFiles, ensure it has lifetime TFL_GhcSession.

addFilesToClean :: DynFlags -> TempFileLifetime -> [FilePath] -> IO ()

Ensure that new_files are cleaned on the next call of cleanTempFiles or cleanCurrentModuleTempFiles, depending on lifetime. If any of new_files are already tracked, they will have their lifetime updated.

changeTempFilesLifetime :: DynFlags -> TempFileLifetime -> [FilePath] -> IO ()

Update the lifetime of files already being tracked. If any files are not being tracked they will be discarded.

newTempName :: DynFlags -> TempFileLifetime -> Suffix -> IO FilePath

newTempLibName :: DynFlags -> TempFileLifetime -> Suffix -> IO (FilePath, FilePath, String)