GHC.ByteCode.Types

Description

Bytecode assembler types

Documentation

data CompiledByteCode

Constructors

CompiledByteCode

Fields

bc_bcos :: FlatBag UnlinkedBCO

Bunch of interpretable bindings

bc_itbls :: ItblEnv

Mapping from DataCons to their info tables

bc_ffis :: [FFIInfo]

ffi blocks we allocated

bc_strs :: AddrEnv

top-level strings (heap allocated)

bc_breaks :: Maybe ModBreaks

breakpoint info (Nothing if breakpoints are disabled)

bc_spt_entries :: ![SptEntry]

Static pointer table entries which should be loaded along with the BCOs. See Note [Grand plan for static forms] in GHC.Iface.Tidy.StaticPtrTable.

seqCompiledByteCode :: CompiledByteCode -> ()

data BCOByteArray a

Wrapper for a ByteArray#. The phantom type tells what elements are stored in the ByteArray#. Creating a ByteArray# can be achieved using UArray's API, where the underlying ByteArray# can be unpacked.

Constructors

BCOByteArray

Fields

getBCOByteArray :: !ByteArray#

mkBCOByteArray :: UArray Int a -> BCOByteArray a

newtype FFIInfo

Constructors

FFIInfo (RemotePtr C_ffi_cif)

newtype RegBitmap

Constructors

RegBitmap

Fields

unRegBitmap :: Word32

data NativeCallType

Constructors

NativePrimCall

NativeTupleReturn

data NativeCallInfo

Constructors

NativeCallInfo

Fields

nativeCallType :: !NativeCallType

nativeCallSize :: !WordOff

nativeCallRegs :: !GlobalRegSet

nativeCallStackSpillSize :: !WordOff

voidTupleReturnInfo :: NativeCallInfo

voidPrimCallInfo :: NativeCallInfo

newtype ByteOff

Constructors

ByteOff Int

newtype WordOff

Constructors

WordOff Int

newtype HalfWord

Constructors

HalfWord Word

data UnlinkedBCO

Constructors

UnlinkedBCO

Fields

unlinkedBCOName :: !Name

unlinkedBCOArity :: !Int

unlinkedBCOInstrs :: !(BCOByteArray Word16)

unlinkedBCOBitmap :: !(BCOByteArray Word)

unlinkedBCOLits :: !(FlatBag BCONPtr)

unlinkedBCOPtrs :: !(FlatBag BCOPtr)

data BCOPtr

Constructors

BCOPtrName !Name

BCOPtrPrimOp !PrimOp

BCOPtrBCO !UnlinkedBCO

BCOPtrBreakArray (ForeignRef BreakArray)
a pointer to a breakpoint's module's BreakArray in GHCi's memory

data BCONPtr

Constructors

BCONPtrWord !Word

BCONPtrLbl !FastString

BCONPtrItbl !Name

BCONPtrAddr !Name
A reference to a top-level string literal; see Note [Generating code for top-level string literal bindings] in GHC.StgToByteCode.

BCONPtrStr !ByteString
Only used internally in the assembler in an intermediate representation; should never appear in a fully-assembled UnlinkedBCO. Also see Note [Allocating string literals] in GHC.ByteCode.Asm.

type ItblEnv = NameEnv (Name, ItblPtr)

newtype ItblPtr

Constructors

ItblPtr (RemotePtr StgInfoTable)

type AddrEnv = NameEnv (Name, AddrPtr)

newtype AddrPtr

Constructors

AddrPtr (RemotePtr ())

data CgBreakInfo

Information about a breakpoint that we know at code-generation time In order to be used, this needs to be hydrated relative to the current HscEnv by hydrateCgBreakInfo. Everything here can be fully forced and that's critical for preventing space leaks (see #22530)

Constructors

CgBreakInfo

Fields

cgb_tyvars :: ![IfaceTvBndr]

Type variables in scope at the breakpoint

cgb_vars :: ![Maybe (IfaceIdBndr, Word)]

cgb_resty :: !IfaceType

data ModBreaks

All the information about the breakpoints for a module

Constructors

ModBreaks

Fields

modBreaks_flags :: ForeignRef BreakArray

The array of flags, one per breakpoint, indicating which breakpoints are enabled.

modBreaks_locs :: !(Array BreakIndex SrcSpan)

An array giving the source span of each breakpoint.

modBreaks_vars :: !(Array BreakIndex [OccName])

An array giving the names of the free variables at each breakpoint.

modBreaks_decls :: !(Array BreakIndex [String])

An array giving the names of the declarations enclosing each breakpoint. See Note [Field modBreaks_decls]

modBreaks_ccs :: !(Array BreakIndex (RemotePtr CostCentre))

Array pointing to cost centre for each breakpoint

modBreaks_breakInfo :: IntMap CgBreakInfo

info about each breakpoint from the bytecode generator

modBreaks_module :: RemotePtr ModuleName

type BreakIndex = Int

Breakpoint index

emptyModBreaks :: ModBreaks

Construct an empty ModBreaks

data CCostCentre

C CostCentre type

data FlatBag a

Store elements in a flattened representation.

A FlatBag is a data structure that stores an ordered list of elements in a flat structure, avoiding the overhead of a linked list. Use this data structure, if the code requires the following properties:

Elements are stored in a long-lived object, and benefit from a flattened representation.

The FlatBag will be traversed but not extended or filtered.

The number of elements should be known.

Sharing of the empty case improves memory behaviour.

A FlagBag aims to have as little overhead as possible to store its elements. To achieve that, it distinguishes between the empty case, singleton, tuple and general case. Thus, we only pay for the additional three words of an Array if we have at least three elements.

sizeFlatBag :: FlatBag a -> Word

Calculate the size of

fromSizedSeq :: SizedSeq a -> FlatBag a

Convert a SizedSeq into its flattened representation. A 'FlatBag a' is more memory efficient than '[a]', if no further modification is necessary.

elemsFlatBag :: FlatBag a -> [a]

Get all elements that are stored in the FlatBag.