GHC.Cmm

Cmm top-level datatypes

type CmmProgram = [CmmGroup]

type CmmGroup = GenCmmGroup CmmStatics CmmTopInfo CmmGraph

Cmm group before SRT generation

type CmmGroupSRTs = GenCmmGroup RawCmmStatics CmmTopInfo CmmGraph

Cmm group with SRTs

type RawCmmGroup = GenCmmGroup RawCmmStatics (LabelMap RawCmmStatics) CmmGraph

Raw cmm group (TODO (osa): not sure what that means)

type GenCmmGroup d h g = [GenCmmDecl d h g]

type CmmDecl = GenCmmDecl CmmStatics CmmTopInfo CmmGraph

type CmmDeclSRTs = GenCmmDecl RawCmmStatics CmmTopInfo CmmGraph

data GenCmmDecl d h g

A top-level chunk, abstracted over the type of the contents of the basic blocks (Cmm or instructions are the likely instantiations).

Constructors

CmmProc h CLabel [GlobalReg] g

CmmData Section d

type CmmGraph = GenCmmGraph CmmNode

data GenCmmGraph n

Constructors

CmmGraph

Fields

g_entry :: BlockId

g_graph :: Graph n C C

type CmmBlock = Block CmmNode C C

type RawCmmDecl = GenCmmDecl RawCmmStatics (LabelMap RawCmmStatics) CmmGraph

data Section

Constructors

Section SectionType CLabel

data SectionType

Constructors

Text

Data

ReadOnlyData

RelocatableReadOnlyData

UninitialisedData

ReadOnlyData16

CString

OtherSection String

data GenCmmStatics (rawOnly :: Bool) where

Constructors

CmmStatics :: CLabel -> CmmInfoTable -> CostCentreStack -> [CmmLit] -> GenCmmStatics 'False

CmmStaticsRaw :: CLabel -> [CmmStatic] -> GenCmmStatics a
Static data, after SRTs are generated

type CmmStatics = GenCmmStatics 'False

type RawCmmStatics = GenCmmStatics 'True

data CmmStatic

Constructors

CmmStaticLit CmmLit
a literal value, size given by cmmLitRep of the literal.

CmmUninitialised Int
uninitialised data, N bytes long

CmmString ByteString
string of 8-bit values only, not zero terminated.

CmmFileEmbed FilePath
an embedded binary file

data SectionProtection

Constructors

ReadWriteSection

ReadOnlySection

WriteProtectedSection

sectionProtection :: Section -> SectionProtection

Should a data in this section be considered constant at runtime

Blocks containing lists

data GenBasicBlock i

Constructors

BasicBlock BlockId [i]

blockId :: GenBasicBlock i -> BlockId

The branch block id is that of the first block in the branch, which is that branch's entry point

newtype ListGraph i

Constructors

ListGraph [GenBasicBlock i]

pprBBlock :: Outputable stmt => GenBasicBlock stmt -> SDoc

Info Tables

data CmmTopInfo

CmmTopInfo is attached to each CmmDecl (see defn of CmmGroup), and contains the extra info (beyond the executable code) that belongs to that CmmDecl.

Constructors

TopInfo

Fields

info_tbls :: LabelMap CmmInfoTable

stack_info :: CmmStackInfo

data CmmStackInfo

Constructors

StackInfo

Fields

arg_space :: ByteOff

do_layout :: Bool

data CmmInfoTable

Info table as a haskell data type

Constructors

CmmInfoTable

Fields

cit_lbl :: CLabel

cit_rep :: SMRep

cit_prof :: ProfilingInfo

cit_srt :: Maybe CLabel

cit_clo :: Maybe (Id, CostCentreStack)

topInfoTable :: GenCmmDecl a CmmTopInfo (GenCmmGraph n) -> Maybe CmmInfoTable

data ClosureTypeInfo

Constructors

Constr ConTagZ ConstrDescription

Fun FunArity ArgDescr

Thunk

ThunkSelector SelectorOffset

BlackHole

IndStatic

data ProfilingInfo

Constructors

NoProfilingInfo

ProfilingInfo ByteString ByteString

type ConstrDescription = ByteString

Statements, expressions and types

module GHC.Cmm.Node

module GHC.Cmm.Expr