GHC.CmmToAsm

Description

Native code generator

The native-code generator has machine-independent and machine-dependent modules.

This module (GHC.CmmToAsm) is the top-level machine-independent module. Before entering machine-dependent land, we do some machine-independent optimisations (defined below) on the CmmStmtss.

We convert to the machine-specific Instr datatype with cmmCodeGen, assuming an infinite supply of registers. We then use a machine-independent register allocator (regAlloc) to rejoin reality. Obviously, regAlloc has machine-specific helper functions (see about RegAllocInfo below).

Finally, we order the basic blocks of the function so as to minimise the number of jumps between blocks, by utilising fallthrough wherever possible.

The machine-dependent bits break down as follows:

[MachRegs] Everything about the target platform's machine registers (and immediate operands, and addresses, which tend to intermingle/interact with registers).

[MachInstrs] Includes the Instr datatype (possibly should have a module of its own), plus a miscellany of other things (e.g., targetDoubleSize, smStablePtrTable, ...)

[MachCodeGen] is where Cmm stuff turns into machine instructions.

[PprMach] pprInstr turns an Instr into text (well, really a SDoc).

[RegAllocInfo] In the register allocator, we manipulate MRegsStates, which are BitSets, one bit per machine register. When we want to say something about a specific machine register (e.g., ``it gets clobbered by this instruction''), we set/unset its bit. Obviously, we do this BitSet thing for efficiency reasons.

The RegAllocInfo module collects together the machine-specific info needed to do register allocation.

[RegisterAlloc] The (machine-independent) register allocator. -}

Documentation

nativeCodeGen :: forall a. Logger -> DynFlags -> Module -> ModLocation -> Handle -> UniqSupply -> Stream IO RawCmmGroup a -> IO a

Test-only exports: see trac #12744

cmmNativeGen

Arguments

:: forall statics instr jumpDest. (Instruction instr, OutputableP Platform statics, Outputable jumpDest)

=> Logger

-> DynFlags

-> ModLocation

-> NcgImpl statics instr jumpDest

-> UniqSupply

-> DwarfFiles

-> LabelMap DebugBlock

-> RawCmmDecl
the cmm to generate code for

-> Int
sequence number of this top thing

-> IO (UniqSupply, DwarfFiles, [NatCmmDecl statics instr], [CLabel], Maybe [RegAllocStats statics instr], Maybe [RegAllocStats], LabelMap [UnwindPoint])

Complete native code generation phase for a single top-level chunk of Cmm. Dumping the output of each stage along the way. Global conflict graph and NGC stats

data NcgImpl statics instr jumpDest

Constructors

NcgImpl

Fields

ncgConfig :: !NCGConfig

cmmTopCodeGen :: RawCmmDecl -> NatM [NatCmmDecl statics instr]

generateJumpTableForInstr :: instr -> Maybe (NatCmmDecl statics instr)

getJumpDestBlockId :: jumpDest -> Maybe BlockId

canShortcut :: instr -> Maybe jumpDest

shortcutStatics :: (BlockId -> Maybe jumpDest) -> statics -> statics

shortcutJump :: (BlockId -> Maybe jumpDest) -> instr -> instr

pprNatCmmDecl :: NatCmmDecl statics instr -> SDoc

Module is only for printing internal labels. See Note [Internal proc labels] in CLabel.

maxSpillSlots :: Int

allocatableRegs :: [RealReg]

ncgExpandTop :: [NatCmmDecl statics instr] -> [NatCmmDecl statics instr]

ncgAllocMoreStack :: Int -> NatCmmDecl statics instr -> UniqSM (NatCmmDecl statics instr, [(BlockId, BlockId)])

The list of block ids records the redirected jumps to allow us to update the CFG.

ncgMakeFarBranches :: LabelMap RawCmmStatics -> [NatBasicBlock instr] -> [NatBasicBlock instr]

extractUnwindPoints :: [instr] -> [UnwindPoint]

given the instruction sequence of a block, produce a list of the block's UnwindPoints See Note [What is this unwinding business?] in GHC.Cmm.DebugBlock and Note [Unwinding information in the NCG] in this module.

invertCondBranches :: Maybe CFG -> LabelMap RawCmmStatics -> [NatBasicBlock instr] -> [NatBasicBlock instr]

Turn the sequence of jcc l1; jmp l2 into jncc l2; <block_l1> when possible.

initNCGConfig :: DynFlags -> Module -> NCGConfig

Initialize the native code generator configuration from the DynFlags