GHC.Core.Coercion.Axiom

Description

Module for coercion axioms, used to represent type family instances and newtypes

Documentation

data BranchFlag

type Branched = 'Branched

type Unbranched = 'Unbranched

type BranchIndex = Int

newtype Branches (br :: BranchFlag)

Constructors

MkBranches

Fields

unMkBranches :: Array BranchIndex CoAxBranch

manyBranches :: [CoAxBranch] -> Branches Branched

unbranched :: CoAxBranch -> Branches Unbranched

fromBranches :: Branches br -> [CoAxBranch]

numBranches :: Branches br -> Int

mapAccumBranches :: ([CoAxBranch] -> CoAxBranch -> CoAxBranch) -> Branches br -> Branches br

The [CoAxBranch] passed into the mapping function is a list of all previous branches, reversed

data CoAxiom br

A CoAxiom is a "coercion constructor", i.e. a named equality axiom.

Constructors

CoAxiom

Fields

co_ax_unique :: Unique

co_ax_name :: Name

co_ax_role :: Role

co_ax_tc :: TyCon

co_ax_branches :: Branches br

co_ax_implicit :: Bool

data CoAxBranch

Constructors

CoAxBranch

Fields

cab_loc :: SrcSpan

cab_tvs :: [TyVar]

cab_eta_tvs :: [TyVar]

cab_cvs :: [CoVar]

cab_roles :: [Role]

cab_lhs :: [Type]

cab_rhs :: Type

cab_incomps :: [CoAxBranch]

toBranchedAxiom :: CoAxiom br -> CoAxiom Branched

toUnbranchedAxiom :: CoAxiom br -> CoAxiom Unbranched

coAxiomName :: CoAxiom br -> Name

coAxiomArity :: CoAxiom br -> BranchIndex -> Arity

coAxiomBranches :: CoAxiom br -> Branches br

coAxiomTyCon :: CoAxiom br -> TyCon

isImplicitCoAxiom :: CoAxiom br -> Bool

coAxiomNumPats :: CoAxiom br -> Int

coAxiomNthBranch :: CoAxiom br -> BranchIndex -> CoAxBranch

coAxiomSingleBranch_maybe :: CoAxiom br -> Maybe CoAxBranch

coAxiomRole :: CoAxiom br -> Role

coAxiomSingleBranch :: CoAxiom Unbranched -> CoAxBranch

coAxBranchTyVars :: CoAxBranch -> [TyVar]

coAxBranchCoVars :: CoAxBranch -> [CoVar]

coAxBranchRoles :: CoAxBranch -> [Role]

coAxBranchLHS :: CoAxBranch -> [Type]

coAxBranchRHS :: CoAxBranch -> Type

coAxBranchSpan :: CoAxBranch -> SrcSpan

coAxBranchIncomps :: CoAxBranch -> [CoAxBranch]

placeHolderIncomps :: [CoAxBranch]

data Role

Constructors

Nominal

Representational

Phantom

fsFromRole :: Role -> FastString

data CoAxiomRule

For now, we work only with nominal equality.

Constructors

CoAxiomRule

Fields

coaxrName :: FastString

coaxrAsmpRoles :: [Role]

coaxrRole :: Role

coaxrProves :: [TypeEqn] -> Maybe TypeEqn

coaxrProves returns Nothing when it doesn't like the supplied arguments. When this happens in a coercion that means that the coercion is ill-formed, and Core Lint checks for that.

type TypeEqn = Pair Type

A more explicit representation for `t1 ~ t2`.

data BuiltInSynFamily

Constructors

BuiltInSynFamily

Fields

sfMatchFam :: [Type] -> Maybe (CoAxiomRule, [Type], Type)

sfInteractTop :: [Type] -> Type -> [TypeEqn]

sfInteractInert :: [Type] -> Type -> [Type] -> Type -> [TypeEqn]

trivialBuiltInFamily :: BuiltInSynFamily