GHC.Core.Lint

Documentation

lintCoreBindings :: DynFlags -> CoreToDo -> [Var] -> CoreProgram -> WarnsAndErrs

Type-check a CoreProgram. See Note [Core Lint guarantee].

lintUnfolding :: Bool -> DynFlags -> SrcLoc -> VarSet -> CoreExpr -> Maybe (Bag SDoc)

lintPassResult :: HscEnv -> CoreToDo -> CoreProgram -> IO ()

lintInteractiveExpr

Arguments

:: SDoc
The source of the linted expression

-> HscEnv

-> CoreExpr

-> IO ()

lintExpr :: DynFlags -> [Var] -> CoreExpr -> Maybe (Bag SDoc)

lintAnnots :: SDoc -> (ModGuts -> CoreM ModGuts) -> ModGuts -> CoreM ModGuts

This checks whether a pass correctly looks through debug annotations (SourceNote). This works a bit different from other consistency checks: We check this by running the given task twice, noting all differences between the results.

lintAxioms

Arguments

:: Logger

-> DynFlags

-> SDoc
The source of the linted axioms

-> [CoAxiom Branched]

-> IO ()

interactiveInScope :: InteractiveContext -> [Var]

Debug output

endPass :: CoreToDo -> CoreProgram -> [CoreRule] -> CoreM ()

endPassIO :: HscEnv -> PrintUnqualified -> CoreToDo -> CoreProgram -> [CoreRule] -> IO ()

displayLintResults

Arguments

:: Logger

-> DynFlags

-> Bool
If True, display linter warnings. If False, ignore linter warnings.

-> SDoc
The source of the linted program

-> SDoc
The linted program, pretty-printed

-> WarnsAndErrs

-> IO ()

dumpPassResult :: Logger -> DynFlags -> PrintUnqualified -> Maybe DumpFlag -> SDoc -> SDoc -> CoreProgram -> [CoreRule] -> IO ()

dumpIfSet :: Logger -> DynFlags -> Bool -> CoreToDo -> SDoc -> SDoc -> IO ()