GHC.Core.Opt.Arity

Description

Arity and eta expansion

Documentation

manifestArity :: CoreExpr -> Arity

manifestArity sees how many leading value lambdas there are, after looking through casts

joinRhsArity :: CoreExpr -> JoinArity

exprArity :: CoreExpr -> Arity

An approximate, fast, version of exprEtaExpandArity

typeArity :: Type -> [OneShotInfo]

exprEtaExpandArity :: DynFlags -> CoreExpr -> ArityType

The Arity returned is the number of value args the expression can be applied to without doing much work

findRhsArity :: DynFlags -> Id -> CoreExpr -> Arity -> ArityType

etaExpand :: Arity -> CoreExpr -> CoreExpr

etaExpand n e returns an expression with the same meaning as e, but with arity n.

Given:

e' = etaExpand n e

We should have that:

ty = exprType e = exprType e'

etaExpandAT :: ArityType -> CoreExpr -> CoreExpr

exprBotStrictness_maybe :: CoreExpr -> Maybe (Arity, StrictSig)

ArityType

data ArityType

The analysis lattice of arity analysis. It is isomorphic to

   data ArityType'
     = AEnd Divergence
     | ALam OneShotInfo ArityType'

Which is easier to display the Hasse diagram for:

 ALam OneShotLam at
         |
     AEnd topDiv
         |
 ALam NoOneShotInfo at
         |
     AEnd exnDiv
         |
     AEnd botDiv

where the at fields of ALam are inductively subject to the same order. That is, ALam os at1 < ALam os at2 iff at1 < at2.

Why the strange Top element? See Note [Combining case branches: optimistic one-shot-ness]

We rely on this lattice structure for fixed-point iteration in findRhsArity. For the semantics of ArityType, see Note [ArityType].

Constructors

AT ![OneShotInfo] !Divergence
AT oss div means this value can safely be eta-expanded length oss times, provided use sites respect the OneShotInfos in oss. A OneShotLam annotation can come from two sources: * The user annotated a lambda as one-shot with oneShot * It's from a lambda binder of a type affected by `-fstate-hack`. See idStateHackOneShotInfo. In both cases, OneShotLam should win over NoOneShotInfo, see Note [Combining case branches].

If div is dead-ending (isDeadEndDiv), then application to length os arguments will surely diverge, similar to the situation with DmdType.

mkBotArityType :: [OneShotInfo] -> ArityType

mkManifestArityType :: [Var] -> CoreExpr -> ArityType

expandableArityType :: ArityType -> Bool

True = eta-expansion will add at least one lambda

arityTypeArity :: ArityType -> Arity

The number of value args for the arity type

maxWithArity :: ArityType -> Arity -> ArityType

Expand a non-bottoming arity type so that it has at least the given arity.

idArityType :: Id -> ArityType

Join points

etaExpandToJoinPoint :: JoinArity -> CoreExpr -> ([CoreBndr], CoreExpr)

Split an expression into the given number of binders and a body, eta-expanding if necessary. Counts value *and* type binders.

etaExpandToJoinPointRule :: JoinArity -> CoreRule -> CoreRule

Coercions and casts

pushCoArg :: CoercionR -> CoreArg -> Maybe (CoreArg, MCoercion)

pushCoArgs :: CoercionR -> [CoreArg] -> Maybe ([CoreArg], MCoercion)

pushCoValArg :: CoercionR -> Maybe (MCoercionR, MCoercionR)

pushCoTyArg :: CoercionR -> Type -> Maybe (Type, MCoercionR)

pushCoercionIntoLambda :: HasDebugCallStack => InScopeSet -> Var -> CoreExpr -> CoercionR -> Maybe (Var, CoreExpr)

pushCoDataCon :: DataCon -> [CoreExpr] -> Coercion -> Maybe (DataCon, [Type], [CoreExpr])

collectBindersPushingCo :: CoreExpr -> ([Var], CoreExpr)