GHC.Core.Opt.ConstantFold

Description

Constant Folder

Documentation

primOpRules :: Name -> PrimOp -> Maybe CoreRule

builtinRules :: [CoreRule]

caseRules :: Platform -> CoreExpr -> Maybe (CoreExpr, AltCon -> Maybe AltCon, Id -> CoreExpr)

Match the scrutinee of a case and potentially return a new scrutinee and a function to apply to each literal alternative.