GHC.Core.Opt.Monad

Configuration of the core-to-core passes

data CoreToDo

Constructors

CoreDoSimplify Int SimplMode

CoreDoPluginPass String CorePluginPass

CoreDoFloatInwards

CoreDoFloatOutwards FloatOutSwitches

CoreLiberateCase

CoreDoPrintCore

CoreDoStaticArgs

CoreDoCallArity

CoreDoExitify

CoreDoDemand

CoreDoCpr

CoreDoWorkerWrapper

CoreDoSpecialising

CoreDoSpecConstr

CoreCSE

CoreDoRuleCheck CompilerPhase String

CoreDoNothing

CoreDoPasses [CoreToDo]

CoreDesugar

CoreDesugarOpt

CoreTidy

CorePrep

CoreAddCallerCcs

CoreOccurAnal

runWhen :: Bool -> CoreToDo -> CoreToDo

runMaybe :: Maybe a -> (a -> CoreToDo) -> CoreToDo

data SimplMode

Constructors

SimplMode

Fields

sm_names :: [String]

Name(s) of the phase

sm_phase :: CompilerPhase

sm_uf_opts :: !UnfoldingOpts

Unfolding options

sm_rules :: !Bool

Whether RULES are enabled

sm_inline :: !Bool

Whether inlining is enabled

sm_case_case :: !Bool

Whether case-of-case is enabled

sm_eta_expand :: !Bool

Whether eta-expansion is enabled

sm_cast_swizzle :: !Bool

Do we swizzle casts past lambdas?

sm_pre_inline :: !Bool

Whether pre-inlining is enabled

sm_logger :: !Logger

sm_dflags :: DynFlags

data FloatOutSwitches

Constructors

FloatOutSwitches

Fields

floatOutLambdas :: Maybe Int

Just n = float lambdas to top level, if doing so will abstract over n or fewer value variables Nothing = float all lambdas to top level, regardless of how many free variables Just 0 is the vanilla case: float a lambda iff it has no free vars

floatOutConstants :: Bool

True = float constants to top level, even if they do not escape a lambda

floatOutOverSatApps :: Bool

True = float out over-saturated applications based on arity information. See Note [Floating over-saturated applications] in GHC.Core.Opt.SetLevels

floatToTopLevelOnly :: Bool

Allow floating to the top level only.

pprPassDetails :: CoreToDo -> SDoc

Plugins

type CorePluginPass = ModGuts -> CoreM ModGuts

A description of the plugin pass itself

bindsOnlyPass :: (CoreProgram -> CoreM CoreProgram) -> ModGuts -> CoreM ModGuts

Counting

data SimplCount

doSimplTick :: DynFlags -> Tick -> SimplCount -> SimplCount

doFreeSimplTick :: Tick -> SimplCount -> SimplCount

simplCountN :: SimplCount -> Int

pprSimplCount :: SimplCount -> SDoc

plusSimplCount :: SimplCount -> SimplCount -> SimplCount

zeroSimplCount :: DynFlags -> SimplCount

isZeroSimplCount :: SimplCount -> Bool

hasDetailedCounts :: SimplCount -> Bool

data Tick

Constructors

PreInlineUnconditionally Id

PostInlineUnconditionally Id

UnfoldingDone Id

RuleFired FastString

LetFloatFromLet

EtaExpansion Id

EtaReduction Id

BetaReduction Id

CaseOfCase Id

KnownBranch Id

CaseMerge Id

AltMerge Id

CaseElim Id

CaseIdentity Id

FillInCaseDefault Id

SimplifierDone

The monad

data CoreM a

The monad used by Core-to-Core passes to register simplification statistics. Also used to have common state (in the form of UniqueSupply) for generating Uniques.

runCoreM

Arguments

:: HscEnv

-> RuleBase

-> Char
Mask

-> Module

-> ModuleSet

-> PrintUnqualified

-> SrcSpan

-> CoreM a

-> IO (a, SimplCount)

Reading from the monad

getHscEnv :: CoreM HscEnv

getRuleBase :: CoreM RuleBase

getModule :: HasModule m => m Module

getDynFlags :: HasDynFlags m => m DynFlags

getPackageFamInstEnv :: CoreM PackageFamInstEnv

getVisibleOrphanMods :: CoreM ModuleSet

getUniqMask :: CoreM Char

getPrintUnqualified :: CoreM PrintUnqualified

getSrcSpanM :: CoreM SrcSpan

Writing to the monad

addSimplCount :: SimplCount -> CoreM ()

Lifting into the monad

liftIO :: MonadIO m => IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3

liftIOWithCount :: IO (SimplCount, a) -> CoreM a

Lift an IO operation into CoreM while consuming its SimplCount

Dealing with annotations

getAnnotations :: Typeable a => ([Word8] -> a) -> ModGuts -> CoreM (ModuleEnv [a], NameEnv [a])

Get all annotations of a given type. This happens lazily, that is no deserialization will take place until the [a] is actually demanded and the [a] can also be empty (the UniqFM is not filtered).

This should be done once at the start of a Core-to-Core pass that uses annotations.

See Note [Annotations]

getFirstAnnotations :: Typeable a => ([Word8] -> a) -> ModGuts -> CoreM (ModuleEnv a, NameEnv a)

Get at most one annotation of a given type per annotatable item.

Screen output

putMsg :: SDoc -> CoreM ()

Output a message to the screen

putMsgS :: String -> CoreM ()

Output a String message to the screen

errorMsg :: SDoc -> CoreM ()

Output an error to the screen. Does not cause the compiler to die.

errorMsgS :: String -> CoreM ()

Output an error to the screen. Does not cause the compiler to die.

warnMsg :: WarnReason -> SDoc -> CoreM ()

fatalErrorMsg :: SDoc -> CoreM ()

Output a fatal error to the screen. Does not cause the compiler to die.

fatalErrorMsgS :: String -> CoreM ()

Output a fatal error to the screen. Does not cause the compiler to die.

debugTraceMsg :: SDoc -> CoreM ()

Outputs a debugging message at verbosity level of -v or higher

debugTraceMsgS :: String -> CoreM ()

Output a string debugging message at verbosity level of -v or higher

dumpIfSet_dyn :: DumpFlag -> String -> DumpFormat -> SDoc -> CoreM ()

Show some labelled SDoc if a particular flag is set or at a verbosity level of -v -ddump-most or higher