GHC.Core.Unfold

Documentation

data Unfolding

Records the unfolding of an identifier, which is approximately the form the identifier would have if we substituted its definition in for the identifier. This type should be treated as abstract everywhere except in GHC.Core.Unfold

data UnfoldingGuidance

UnfoldingGuidance says when unfolding should take place

data UnfoldingOpts

Unfolding options

Constructors

UnfoldingOpts

Fields

unfoldingCreationThreshold :: !Int

Threshold above which unfoldings are not *created*

unfoldingUseThreshold :: !Int

Threshold above which unfoldings are not *inlined*

unfoldingFunAppDiscount :: !Int

Discount for lambdas that are used (applied)

unfoldingDictDiscount :: !Int

Discount for dictionaries

unfoldingVeryAggressive :: !Bool

Force inlining in many more cases

unfoldingCaseThreshold :: !Int

unfoldingCaseScaling :: !Int

defaultUnfoldingOpts :: UnfoldingOpts

updateCreationThreshold :: Int -> UnfoldingOpts -> UnfoldingOpts

updateUseThreshold :: Int -> UnfoldingOpts -> UnfoldingOpts

updateFunAppDiscount :: Int -> UnfoldingOpts -> UnfoldingOpts

updateDictDiscount :: Int -> UnfoldingOpts -> UnfoldingOpts

updateVeryAggressive :: Bool -> UnfoldingOpts -> UnfoldingOpts

updateCaseScaling :: Int -> UnfoldingOpts -> UnfoldingOpts

updateCaseThreshold :: Int -> UnfoldingOpts -> UnfoldingOpts

data ArgSummary

Constructors

TrivArg

NonTrivArg

ValueArg

couldBeSmallEnoughToInline :: UnfoldingOpts -> Int -> CoreExpr -> Bool

inlineBoringOk :: CoreExpr -> Bool

certainlyWillInline :: UnfoldingOpts -> IdInfo -> Maybe Unfolding

Sees if the unfolding is pretty certain to inline. If so, return a *stable* unfolding for it, that will always inline.

smallEnoughToInline :: UnfoldingOpts -> Unfolding -> Bool

callSiteInline :: Logger -> DynFlags -> Int -> Id -> Bool -> Bool -> [ArgSummary] -> CallCtxt -> Maybe CoreExpr

data CallCtxt

Constructors

BoringCtxt

RhsCtxt

DiscArgCtxt

RuleArgCtxt

ValAppCtxt

CaseCtxt

calcUnfoldingGuidance :: UnfoldingOpts -> Bool -> CoreExpr -> UnfoldingGuidance