GHC.Driver.CmdLine

Description

Command-line parser

This is an abstract command-line parser used by DynFlags.

(c) The University of Glasgow 2005

Documentation

processArgs

Arguments

:: Monad m

=> [Flag m]
cmdline parser spec

-> [Located String]
args

-> (FilePath -> EwM m [Located String])
response file handler

-> m ([Located String], [Err], [Warn])

parseResponseFile :: MonadIO m => FilePath -> EwM m [Located String]

Parse a response file into arguments.

data OptKind m

Constructors

NoArg (EwM m ())

HasArg (String -> EwM m ())

SepArg (String -> EwM m ())

Prefix (String -> EwM m ())

OptPrefix (String -> EwM m ())

OptIntSuffix (Maybe Int -> EwM m ())

IntSuffix (Int -> EwM m ())

WordSuffix (Word -> EwM m ())

FloatSuffix (Float -> EwM m ())

PassFlag (String -> EwM m ())

AnySuffix (String -> EwM m ())

data GhcFlagMode

GHC flag modes describing when a flag has an effect.

Constructors

OnlyGhc
The flag only affects the non-interactive GHC

OnlyGhci
The flag only affects the interactive GHC

AllModes
The flag affects multiple ghc modes

HiddenFlag
This flag should not be seen in cli completion

data Flag m

Constructors

Flag

Fields

flagName :: String

flagOptKind :: OptKind m

flagGhcMode :: GhcFlagMode

defFlag :: String -> OptKind m -> Flag m

defGhcFlag :: String -> OptKind m -> Flag m

defGhciFlag :: String -> OptKind m -> Flag m

defHiddenFlag :: String -> OptKind m -> Flag m

hoistFlag :: forall m n. (forall a. m a -> n a) -> Flag m -> Flag n

errorsToGhcException :: [(String, String)] -> GhcException

newtype Err

A command-line error message

Constructors

Err

Fields

errMsg :: Located String

data Warn

A command-line warning message and the reason it arose

Constructors

Warn

Fields

warnReason :: DiagnosticReason

warnMsg :: Located String

data WarnReason

Used when filtering warnings: if a reason is given it can be filtered out when displaying.

Constructors

NoReason

ReasonDeprecatedFlag

ReasonUnrecognisedFlag

data EwM m a

runEwM :: EwM m a -> m (Errs, Warns, a)

addErr :: Monad m => String -> EwM m ()

addWarn :: Monad m => String -> EwM m ()

addFlagWarn :: Monad m => DiagnosticReason -> String -> EwM m ()

getArg :: Monad m => EwM m String

getCurLoc :: Monad m => EwM m SrcSpan

liftEwM :: Monad m => m a -> EwM m a