GHC.Driver.Pipeline

Documentation

oneShot :: HscEnv -> Phase -> [(String, Maybe Phase)] -> IO ()

compileFile :: HscEnv -> Phase -> (FilePath, Maybe Phase) -> IO FilePath

preprocess

Arguments

:: HscEnv

-> FilePath
input filename

-> Maybe InputFileBuffer
optional buffer to use instead of reading the input file

-> Maybe Phase
starting phase

-> IO (Either ErrorMessages (DynFlags, FilePath))

Just preprocess a file, put the result in a temp. file (used by the compilation manager during the summary phase).

We return the augmented DynFlags, because they contain the result of slurping in the OPTIONS pragmas

compileOne

Arguments

:: HscEnv

-> ModSummary
summary for module being compiled

-> Int
module N ...

-> Int
... of M

-> Maybe ModIface
old interface, if we have one

-> Maybe Linkable
old linkable, if we have one

-> SourceModified

-> IO HomeModInfo
the complete HomeModInfo, if successful

Compile

Compile a single module, under the control of the compilation manager.

This is the interface between the compilation manager and the compiler proper (hsc), where we deal with tedious details like reading the OPTIONS pragma from the source file, converting the C or assembly that GHC produces into an object file, and compiling FFI stub files.

NB. No old interface can also mean that the source has changed.

compileOne'

Arguments

:: Maybe TcGblEnv

-> Maybe Messager

-> HscEnv

-> ModSummary
summary for module being compiled

-> Int
module N ...

-> Int
... of M

-> Maybe ModIface
old interface, if we have one

-> Maybe Linkable
old linkable, if we have one

-> SourceModified

-> IO HomeModInfo
the complete HomeModInfo, if successful

link

Arguments

:: GhcLink
interactive or batch

-> Logger
Logger

-> TmpFs

-> Hooks

-> DynFlags
dynamic flags

-> UnitEnv
unit environment

-> Bool
attempt linking in batch mode?

-> HomePackageTable
what to link

-> IO SuccessFlag

data PhasePlus

Constructors

RealPhase Phase

HscOut HscSource ModuleName HscStatus

newtype CompPipeline a

Constructors

P

Fields

unP :: PipeEnv -> PipeState -> IO (PipeState, a)

data PipeEnv

Constructors

PipeEnv

Fields

stop_phase :: Phase

Stop just before this phase

src_filename :: String

basename of original input source

src_basename :: String

basename of original input source

src_suffix :: String

its extension

output_spec :: PipelineOutput

says where to put the pipeline output

data PipeState

Constructors

PipeState

Fields

hsc_env :: HscEnv

only the DynFlags and the Plugins change in the HscEnv. The DynFlags change at various points, for example when we read the OPTIONS_GHC pragmas in the Cpp phase.

maybe_loc :: Maybe ModLocation

the ModLocation. This is discovered during compilation, in the Hsc phase where we read the module header.

foreign_os :: [FilePath]

additional object files resulting from compiling foreign code. They come from two sources: foreign stubs, and add{C,Cxx,Objc,Objcxx}File from template haskell

iface :: Maybe ModIface

Interface generated by HscOut phase. Only available after the phase runs.

phaseOutputFilename :: Phase -> CompPipeline FilePath

Computes the next output filename after we run next_phase. Like getOutputFilename, but it operates in the CompPipeline monad (which specifies all of the ambient information.)

getOutputFilename :: Logger -> TmpFs -> Phase -> PipelineOutput -> String -> DynFlags -> Phase -> Maybe ModLocation -> IO FilePath

Computes the next output filename for something in the compilation pipeline. This is controlled by several variables:

Phase: the last phase to be run (e.g. stopPhase). This is used to tell if we're in the last phase or not, because in that case flags like -o may be important.

PipelineOutput: is this intended to be a Temporary or Persistent build output? Temporary files just go in a fresh temporary name.

String: what was the basename of the original input file?

DynFlags: the obvious thing

Phase: the phase we want to determine the output filename of.

Maybe ModLocation: the ModLocation of the module we're compiling; this can be used to override the default output of an object file. (TODO: do we actually need this?)

getPipeState :: CompPipeline PipeState

getPipeEnv :: CompPipeline PipeEnv

hscPostBackendPhase :: HscSource -> Backend -> Phase

What phase to run after one of the backend code generators has run

getLocation :: HscSource -> ModuleName -> CompPipeline ModLocation

setModLocation :: ModLocation -> CompPipeline ()

setDynFlags :: DynFlags -> CompPipeline ()

runPhase

Arguments

:: PhasePlus
Run this phase

-> FilePath
name of the input file

-> CompPipeline (PhasePlus, FilePath)

Each phase in the pipeline returns the next phase to execute, and the name of the file in which the output was placed.

We must do things dynamically this way, because we often don't know what the rest of the phases will be until part-way through the compilation: for example, an {-# OPTIONS -fasm #-} at the beginning of a source file can change the latter stages of the pipeline from taking the LLVM route to using the native code generator.

doCpp :: Logger -> TmpFs -> DynFlags -> UnitEnv -> Bool -> FilePath -> FilePath -> IO ()

Run CPP

UnitState is needed to compute MIN_VERSION macros

linkingNeeded :: Logger -> DynFlags -> UnitEnv -> Bool -> [Linkable] -> [UnitId] -> IO Bool

checkLinkInfo :: Logger -> DynFlags -> UnitEnv -> [UnitId] -> FilePath -> IO Bool

writeInterfaceOnlyMode :: DynFlags -> Bool