GHC.Hs

Documentation

module GHC.Hs.Binds

module GHC.Hs.Decls

module GHC.Hs.Expr

module GHC.Hs.ImpExp

module GHC.Hs.Lit

module GHC.Hs.Pat

module GHC.Hs.Types

module GHC.Hs.Utils

module GHC.Hs.Doc

module GHC.Hs.PlaceHolder

module GHC.Hs.Extension

data Fixity

data HsModule pass

Haskell Module

All we actually declare here is the top-level structure for a module.

Constructors

HsModule
AnnKeywordIds

AnnModule,AnnWhere

AnnOpen,AnnSemi, AnnClose for explicit braces and semi around hsmodImports,hsmodDecls if this style is used.

Fields

hsmodName :: Maybe (Located ModuleName)

Nothing: "module X where" is omitted (in which case the next field is Nothing too)

hsmodExports :: Maybe (Located [LIE pass])

Export list

Nothing: export list omitted, so export everything

Just []: export nothing

Just [...]: as you would expect...

AnnKeywordIds : AnnOpen ,AnnClose

hsmodImports :: [LImportDecl pass]

We snaffle interesting stuff out of the imported interfaces early on, adding that info to TyDecls/etc; so this list is often empty, downstream.

hsmodDecls :: [LHsDecl pass]

Type, class, value, and interface signature decls

hsmodDeprecMessage :: Maybe (Located WarningTxt)

reason/explanation for warning/deprecation of this module

AnnKeywordIds : AnnOpen ,AnnClose

hsmodHaddockModHeader :: Maybe LHsDocString

Haddock module info and description, unparsed

AnnKeywordIds : AnnOpen ,AnnClose