GHC.Hs.Binds

Documentation

module Language.Haskell.Syntax.Binds

data NHsValBindsLR idL

Constructors

NValBinds [(RecFlag, LHsBinds idL)] [LSig GhcRn]

data AbsBinds

Typechecked, generalised bindings, used in the output to the type checker. See Note [AbsBinds].

Constructors

AbsBinds

Fields

abs_tvs :: [TyVar]

abs_ev_vars :: [EvVar]

Includes equality constraints

abs_exports :: [ABExport]

AbsBinds only gets used when idL = idR after renaming, but these need to be idL's for the collect... code in HsUtil to have the right type

abs_ev_binds :: [TcEvBinds]

Evidence bindings Why a list? See GHC.Tc.TyCl.Instance Note [Typechecking plan for instance declarations]

abs_binds :: LHsBinds GhcTc

Typechecked user bindings

abs_sig :: Bool

data ABExport

Abstraction Bindings Export

Constructors

ABE

Fields

abe_poly :: Id

Any INLINE pragma is attached to this Id

abe_mono :: Id

abe_wrap :: HsWrapper

See Note [ABExport wrapper] Shape: (forall abs_tvs. abs_ev_vars => abe_mono) ~ abe_poly

abe_prags :: TcSpecPrags

SPECIALISE pragmas

data TcSpecPrags

Type checker Specialisation Pragmas

TcSpecPrags conveys SPECIALISE pragmas from the type checker to the desugarer

Constructors

IsDefaultMethod
Super-specialised: a default method should be macro-expanded at every call site

SpecPrags [LTcSpecPrag]

data AnnSig

Constructors

AnnSig

Fields

asDcolon :: AddEpAnn

asRest :: [AddEpAnn]

newtype IdSig

A type signature in generated code, notably the code generated for record selectors. We simply record the desired Id itself, replete with its name, type and IdDetails. Otherwise it's just like a type signature: there should be an accompanying binding

Constructors

IdSig

Fields

unIdSig :: Id

type LTcSpecPrag = Located TcSpecPrag

Located Type checker Specification Pragmas

data TcSpecPrag

Type checker Specification Pragma

Constructors

SpecPrag Id HsWrapper InlinePragma
The Id to be specialised, a wrapper that specialises the polymorphic function, and inlining spec for the specialised function

ppr_sig :: forall p. OutputableBndrId p => Sig (GhcPass p) -> SDoc

pprDeclList :: [SDoc] -> SDoc

pprLHsBindsForUser :: (OutputableBndrId idL, OutputableBndrId idR, OutputableBndrId id2) => LHsBindsLR (GhcPass idL) (GhcPass idR) -> [LSig (GhcPass id2)] -> [SDoc]

pprLHsBinds :: (OutputableBndrId idL, OutputableBndrId idR) => LHsBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

isEmptyLHsBinds :: LHsBindsLR (GhcPass idL) idR -> Bool

emptyLocalBinds :: HsLocalBindsLR (GhcPass a) (GhcPass b)

eqEmptyLocalBinds :: HsLocalBindsLR a b -> Bool

isEmptyValBinds :: HsValBindsLR (GhcPass a) (GhcPass b) -> Bool

emptyValBindsIn :: HsValBindsLR (GhcPass a) (GhcPass b)

emptyValBindsOut :: HsValBindsLR (GhcPass a) (GhcPass b)

emptyLHsBinds :: LHsBindsLR (GhcPass idL) idR

plusHsValBinds :: HsValBinds (GhcPass a) -> HsValBinds (GhcPass a) -> HsValBinds (GhcPass a)

ppr_monobind :: forall idL idR. (OutputableBndrId idL, OutputableBndrId idR) => HsBindLR (GhcPass idL) (GhcPass idR) -> SDoc

pprTicks :: SDoc -> SDoc -> SDoc

pprTcSpecPrags :: TcSpecPrags -> SDoc

isEmptyIPBindsPR :: HsIPBinds (GhcPass p) -> Bool

isEmptyIPBindsTc :: HsIPBinds GhcTc -> Bool

noSpecPrags :: TcSpecPrags

hasSpecPrags :: TcSpecPrags -> Bool

isDefaultMethod :: TcSpecPrags -> Bool

pprVarSig :: OutputableBndr id => [id] -> SDoc -> SDoc

pragSrcBrackets :: SourceText -> String -> SDoc -> SDoc

Using SourceText in case the pragma was spelled differently or used mixed case

pprSpec :: OutputableBndr id => id -> SDoc -> InlinePragma -> SDoc

extractSpecPragName :: SourceText -> String

Extracts the name for a SPECIALIZE instance pragma. In hsSigDoc, the src field of SpecInstSig signature contains the SourceText for a SPECIALIZE instance pragma of the form: "SourceText {-# SPECIALIZE"

Extraction ensures that all variants of the pragma name (with a Z or an S) are output exactly as used in the pragma.

pprMinimalSig :: OutputableBndr name => LBooleanFormula (GenLocated l name) -> SDoc

hsSigDoc :: forall p. IsPass p => Sig (GhcPass p) -> SDoc

pragBrackets :: SDoc -> SDoc

Orphan instances

OutputableBndrId p => Outputable (FixitySig (GhcPass p))

Instance details

Methods

ppr :: FixitySig (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsIPBinds (GhcPass p))

Instance details

Methods

ppr :: HsIPBinds (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (IPBind (GhcPass p))

Instance details

Methods

ppr :: IPBind (GhcPass p) -> SDoc

Outputable (XRec a RdrName) => Outputable (RecordPatSynField a)

Instance details

Methods

ppr :: RecordPatSynField a -> SDoc

OutputableBndrId p => Outputable (Sig (GhcPass p))

Instance details

Methods

ppr :: Sig (GhcPass p) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr) => Outputable (HsBindLR (GhcPass pl) (GhcPass pr))

Instance details

Methods

ppr :: HsBindLR (GhcPass pl) (GhcPass pr) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr) => Outputable (HsLocalBindsLR (GhcPass pl) (GhcPass pr))

Instance details

Methods

ppr :: HsLocalBindsLR (GhcPass pl) (GhcPass pr) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr) => Outputable (HsValBindsLR (GhcPass pl) (GhcPass pr))

Instance details

Methods

ppr :: HsValBindsLR (GhcPass pl) (GhcPass pr) -> SDoc

(OutputableBndrId l, OutputableBndrId r) => Outputable (PatSynBind (GhcPass l) (GhcPass r))

Instance details

Methods

ppr :: PatSynBind (GhcPass l) (GhcPass r) -> SDoc