GHC.Hs.Decls

Description

Abstract syntax of global declarations.

Definitions for: SynDecl and ConDecl, ClassDecl, InstDecl, DefaultDecl and ForeignDecl.

Toplevel declarations

data HsDecl p

A Haskell Declaration

Constructors

TyClD (XTyClD p) (TyClDecl p)
Type or Class Declaration

InstD (XInstD p) (InstDecl p)
Instance declaration

DerivD (XDerivD p) (DerivDecl p)
Deriving declaration

ValD (XValD p) (HsBind p)
Value declaration

SigD (XSigD p) (Sig p)
Signature declaration

KindSigD (XKindSigD p) (StandaloneKindSig p)
Standalone kind signature

DefD (XDefD p) (DefaultDecl p)
'default' declaration

ForD (XForD p) (ForeignDecl p)
Foreign declaration

WarningD (XWarningD p) (WarnDecls p)
Warning declaration

AnnD (XAnnD p) (AnnDecl p)
Annotation declaration

RuleD (XRuleD p) (RuleDecls p)
Rule declaration

SpliceD (XSpliceD p) (SpliceDecl p)
Splice declaration (Includes quasi-quotes)

DocD (XDocD p) DocDecl
Documentation comment declaration

RoleAnnotD (XRoleAnnotD p) (RoleAnnotDecl p)
Role annotation declaration

XHsDecl !(XXHsDecl p)

type LHsDecl p

Arguments

= XRec p (HsDecl p)
When in a list this may have

AnnKeywordId : AnnSemi

data HsDataDefn pass

Haskell Data type Definition

Constructors

HsDataDefn
Declares a data type or newtype, giving its constructors data/newtype T a = constrs data/newtype instance T [a] = constrs

Fields

dd_ext :: XCHsDataDefn pass

dd_ND :: NewOrData

dd_ctxt :: Maybe (LHsContext pass)

Context

dd_cType :: Maybe (XRec pass CType)

dd_kindSig :: Maybe (LHsKind pass)

Optional kind signature.

(Just k) for a GADT-style data, or data instance decl, with explicit kind sig

Always Nothing for H98-syntax decls

dd_cons :: [LConDecl pass]

Data constructors

For data T a = T1 | T2 a the LConDecls all have ConDeclH98. For data T a where { T1 :: T a } the LConDecls all have ConDeclGADT.

dd_derivs :: HsDeriving pass

Optional 'deriving' clause

XHsDataDefn !(XXHsDataDefn pass)

type HsDeriving pass

Arguments

= [LHsDerivingClause pass]
The optional deriving clauses of a data declaration. Clauses is plural because one can specify multiple deriving clauses using the -XDerivingStrategies language extension.

The list of LHsDerivingClauses corresponds to exactly what the user requested to derive, in order. If no deriving clauses were specified, the list is empty.

Haskell Deriving clause

type LHsFunDep pass = XRec pass (FunDep pass)

data HsDerivingClause pass

A single deriving clause of a data declaration.

AnnKeywordId : AnnDeriving, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

Constructors

HsDerivingClause

Fields

deriv_clause_ext :: XCHsDerivingClause pass

deriv_clause_strategy :: Maybe (LDerivStrategy pass)

The user-specified strategy (if any) to use when deriving deriv_clause_tys.

deriv_clause_tys :: LDerivClauseTys pass

The types to derive.

XHsDerivingClause !(XXHsDerivingClause pass)

type LHsDerivingClause pass = XRec pass (HsDerivingClause pass)

data DerivClauseTys pass

The types mentioned in a single deriving clause. This can come in two forms, DctSingle or DctMulti, depending on whether the types are surrounded by enclosing parentheses or not. These parentheses are semantically different than HsParTy. For example, deriving () means "derive zero classes" rather than "derive an instance of the 0-tuple".

DerivClauseTys use LHsSigType because deriving clauses can mention type variables that aren't bound by the datatype, e.g.

data T b = ... deriving (C [a])

should produce a derived instance for C [a] (T b).

Constructors

DctSingle (XDctSingle pass) (LHsSigType pass)
A deriving clause with a single type. Moreover, that type can only be a type constructor without any arguments.

Example: deriving Eq

DctMulti (XDctMulti pass) [LHsSigType pass]
A deriving clause with a comma-separated list of types, surrounded by enclosing parentheses.

Example: deriving (Eq, C a)

XDerivClauseTys !(XXDerivClauseTys pass)

type LDerivClauseTys pass = XRec pass (DerivClauseTys pass)

data NewOrData

Constructors

NewType
newtype Blah ...

DataType
data Blah ...

newOrDataToFlavour :: NewOrData -> TyConFlavour

Convert a NewOrData to a TyConFlavour

data StandaloneKindSig pass

Constructors

StandaloneKindSig (XStandaloneKindSig pass) (LIdP pass) (LHsSigType pass)

XStandaloneKindSig !(XXStandaloneKindSig pass)

type LStandaloneKindSig pass = XRec pass (StandaloneKindSig pass)

Located Standalone Kind Signature

standaloneKindSigName :: StandaloneKindSig (GhcPass p) -> IdP (GhcPass p)

Class or type declarations

data TyClDecl pass

A type or class declaration.

Constructors

FamDecl
type/data family T :: *->*

AnnKeywordId : AnnType, AnnData, AnnFamily,AnnDcolon, AnnWhere,AnnOpenP, AnnDcolon,AnnCloseP, AnnEqual,AnnRarrow, AnnVbar

Fields

tcdFExt :: XFamDecl pass

tcdFam :: FamilyDecl pass

SynDecl
type declaration

AnnKeywordId : AnnType, AnnEqual,

Fields

tcdSExt :: XSynDecl pass

Post renameer, FVs

tcdLName :: LIdP pass

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdRhs :: LHsType pass

RHS of type declaration

DataDecl
data declaration

AnnKeywordId : AnnData, AnnFamily, AnnNewType, AnnNewType,AnnDcolon AnnWhere,

Fields

tcdDExt :: XDataDecl pass

Post renamer, CUSK flag, FVs

tcdLName :: LIdP pass

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdDataDefn :: HsDataDefn pass

ClassDecl

Fields

tcdCExt :: XClassDecl pass

Post renamer, FVs

tcdCtxt :: Maybe (LHsContext pass)

Context...

tcdLName :: LIdP pass

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdFDs :: [LHsFunDep pass]

Functional deps

tcdSigs :: [LSig pass]

Methods' signatures

tcdMeths :: LHsBinds pass

Default methods

tcdATs :: [LFamilyDecl pass]

Associated types;

tcdATDefs :: [LTyFamDefltDecl pass]

Associated type defaults

tcdDocs :: [LDocDecl pass]

Haddock docs

XTyClDecl !(XXTyClDecl pass)

type LTyClDecl pass = XRec pass (TyClDecl pass)

Located Declaration of a Type or Class

data DataDeclRn

Constructors

DataDeclRn

Fields

tcdDataCusk :: Bool

does this have a CUSK? See Note [CUSKs: complete user-supplied kind signatures]

tcdFVs :: NameSet

data TyClGroup pass

Type or Class Group

Constructors

TyClGroup

Fields

group_ext :: XCTyClGroup pass

group_tyclds :: [LTyClDecl pass]

group_roles :: [LRoleAnnotDecl pass]

group_kisigs :: [LStandaloneKindSig pass]

group_instds :: [LInstDecl pass]

XTyClGroup !(XXTyClGroup pass)

tyClGroupTyClDecls :: [TyClGroup pass] -> [LTyClDecl pass]

tyClGroupInstDecls :: [TyClGroup pass] -> [LInstDecl pass]

tyClGroupRoleDecls :: [TyClGroup pass] -> [LRoleAnnotDecl pass]

tyClGroupKindSigs :: [TyClGroup pass] -> [LStandaloneKindSig pass]

isClassDecl :: TyClDecl pass -> Bool

type class

isDataDecl :: TyClDecl pass -> Bool

True = argument is a data/newtype declaration.

isSynDecl :: TyClDecl pass -> Bool

type or type instance declaration

tcdName :: Anno (IdGhcP p) ~ SrcSpanAnnN => TyClDecl (GhcPass p) -> IdP (GhcPass p)

isFamilyDecl :: TyClDecl pass -> Bool

type/data family declaration

isTypeFamilyDecl :: TyClDecl pass -> Bool

type family declaration

isDataFamilyDecl :: TyClDecl pass -> Bool

data family declaration

isOpenTypeFamilyInfo :: FamilyInfo pass -> Bool

open type family info

isClosedTypeFamilyInfo :: FamilyInfo pass -> Bool

closed type family info

tyFamInstDeclName :: Anno (IdGhcP p) ~ SrcSpanAnnN => TyFamInstDecl (GhcPass p) -> IdP (GhcPass p)

tyFamInstDeclLName :: Anno (IdGhcP p) ~ SrcSpanAnnN => TyFamInstDecl (GhcPass p) -> LocatedN (IdP (GhcPass p))

countTyClDecls :: [TyClDecl pass] -> (Int, Int, Int, Int, Int)

pprTyClDeclFlavour :: TyClDecl (GhcPass p) -> SDoc

tyClDeclLName :: Anno (IdGhcP p) ~ SrcSpanAnnN => TyClDecl (GhcPass p) -> LocatedN (IdP (GhcPass p))

tyClDeclTyVars :: TyClDecl pass -> LHsQTyVars pass

hsDeclHasCusk :: TyClDecl GhcRn -> Bool

Does this declaration have a complete, user-supplied kind signature? See Note [CUSKs: complete user-supplied kind signatures]

famResultKindSignature :: FamilyResultSig (GhcPass p) -> Maybe (LHsKind (GhcPass p))

data FamilyDecl pass

type Family Declaration

Constructors

FamilyDecl

Fields

fdExt :: XCFamilyDecl pass

fdInfo :: FamilyInfo pass

fdTopLevel :: TopLevelFlag

fdLName :: LIdP pass

fdTyVars :: LHsQTyVars pass

fdFixity :: LexicalFixity

fdResultSig :: LFamilyResultSig pass

fdInjectivityAnn :: Maybe (LInjectivityAnn pass)

XFamilyDecl !(XXFamilyDecl pass)

AnnKeywordId : AnnType, AnnData, AnnFamily, AnnWhere, AnnOpenP, AnnDcolon, AnnCloseP, AnnEqual, AnnRarrow, AnnVbar

type LFamilyDecl pass = XRec pass (FamilyDecl pass)

Located type Family Declaration

data FunDep pass

Constructors

FunDep (XCFunDep pass) [LIdP pass] [LIdP pass]

XFunDep !(XXFunDep pass)

Instance declarations

data InstDecl pass

Instance Declaration

Constructors

ClsInstD

Fields

cid_d_ext :: XClsInstD pass

cid_inst :: ClsInstDecl pass

DataFamInstD

Fields

dfid_ext :: XDataFamInstD pass

dfid_inst :: DataFamInstDecl pass

TyFamInstD

Fields

tfid_ext :: XTyFamInstD pass

tfid_inst :: TyFamInstDecl pass

XInstDecl !(XXInstDecl pass)

type LInstDecl pass = XRec pass (InstDecl pass)

Located Instance Declaration

data FamilyInfo pass

Constructors

DataFamily

OpenTypeFamily

ClosedTypeFamily (Maybe [LTyFamInstEqn pass])
Nothing if we're in an hs-boot file and the user said "type family Foo x where .."

data TyFamInstDecl pass

Type Family Instance Declaration

Constructors

TyFamInstDecl

AnnKeywordId : AnnType, AnnInstance,

Fields

tfid_xtn :: XCTyFamInstDecl pass

tfid_eqn :: TyFamInstEqn pass

XTyFamInstDecl !(XXTyFamInstDecl pass)

type LTyFamInstDecl pass = XRec pass (TyFamInstDecl pass)

Located Type Family Instance Declaration

instDeclDataFamInsts :: [LInstDecl (GhcPass p)] -> [DataFamInstDecl (GhcPass p)]

type TyFamDefltDecl = TyFamInstDecl

Type family default declarations. A convenient synonym for TyFamInstDecl. See Note [Type family instance declarations in HsSyn].

type LTyFamDefltDecl pass = XRec pass (TyFamDefltDecl pass)

Located type family default declarations.

newtype DataFamInstDecl pass

Data Family Instance Declaration

Constructors

DataFamInstDecl

AnnKeywordId : AnnData, AnnNewType,AnnInstance, AnnDcolon AnnWhere,AnnOpen, AnnClose

Fields

dfid_eqn :: FamEqn pass (HsDataDefn pass)

type LDataFamInstDecl pass = XRec pass (DataFamInstDecl pass)

Located Data Family Instance Declaration

pprDataFamInstFlavour :: DataFamInstDecl (GhcPass p) -> SDoc

pprTyFamInstDecl :: OutputableBndrId p => TopLevelFlag -> TyFamInstDecl (GhcPass p) -> SDoc

pprHsFamInstLHS :: OutputableBndrId p => IdP (GhcPass p) -> HsOuterFamEqnTyVarBndrs (GhcPass p) -> HsTyPats (GhcPass p) -> LexicalFixity -> Maybe (LHsContext (GhcPass p)) -> SDoc

data FamEqn pass rhs

Family Equation

One equation in a type family instance declaration, data family instance declaration, or type family default. See Note [Type family instance declarations in HsSyn] See Note [Family instance declaration binders]

Constructors

FamEqn

Fields

feqn_ext :: XCFamEqn pass rhs

feqn_tycon :: LIdP pass

feqn_bndrs :: HsOuterFamEqnTyVarBndrs pass

Optional quantified type vars

feqn_pats :: HsTyPats pass

feqn_fixity :: LexicalFixity

Fixity used in the declaration

feqn_rhs :: rhs

AnnKeywordId : AnnEqual

XFamEqn !(XXFamEqn pass rhs)

type TyFamInstEqn pass = FamEqn pass (LHsType pass)

Type Family Instance Equation

type LTyFamInstEqn pass

Arguments

= XRec pass (TyFamInstEqn pass)
May have AnnKeywordId : AnnSemi when in a list

Located Type Family Instance Equation

type HsTyPats pass = [LHsTypeArg pass]

Haskell Type Patterns

type LClsInstDecl pass = XRec pass (ClsInstDecl pass)

Located Class Instance Declaration

data ClsInstDecl pass

Class Instance Declaration

Constructors

ClsInstDecl

Fields

cid_ext :: XCClsInstDecl pass

cid_poly_ty :: LHsSigType pass

cid_binds :: LHsBinds pass

cid_sigs :: [LSig pass]

cid_tyfam_insts :: [LTyFamInstDecl pass]

cid_datafam_insts :: [LDataFamInstDecl pass]

cid_overlap_mode :: Maybe (XRec pass OverlapMode)

AnnKeywordId : AnnOpen, AnnClose,

XClsInstDecl !(XXClsInstDecl pass)

Standalone deriving declarations

data DerivDecl pass

Stand-alone 'deriving instance' declaration

Constructors

DerivDecl

Fields

deriv_ext :: XCDerivDecl pass

deriv_type :: LHsSigWcType pass

The instance type to derive.

It uses an LHsSigWcType because the context is allowed to be a single wildcard:

deriving instance _ => Eq (Foo a)

Which signifies that the context should be inferred.

deriv_strategy :: Maybe (LDerivStrategy pass)

deriv_overlap_mode :: Maybe (XRec pass OverlapMode)

AnnKeywordId : AnnDeriving, AnnInstance, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

XDerivDecl !(XXDerivDecl pass)

type LDerivDecl pass = XRec pass (DerivDecl pass)

Located stand-alone 'deriving instance' declaration

Deriving strategies

data DerivStrategy pass

Which technique the user explicitly requested when deriving an instance.

Constructors

StockStrategy (XStockStrategy pass)
GHC's "standard" strategy, which is to implement a custom instance for the data type. This only works for certain types that GHC knows about (e.g., Eq, Show, Functor when -XDeriveFunctor is enabled, etc.)

AnyclassStrategy (XAnyClassStrategy pass)
-XDeriveAnyClass

NewtypeStrategy (XNewtypeStrategy pass)
-XGeneralizedNewtypeDeriving

ViaStrategy (XViaStrategy pass)
-XDerivingVia

type LDerivStrategy pass = XRec pass (DerivStrategy pass)

A Located DerivStrategy.

derivStrategyName :: DerivStrategy a -> SDoc

A short description of a DerivStrategy'.

foldDerivStrategy :: p ~ GhcPass pass => r -> (XViaStrategy p -> r) -> DerivStrategy p -> r

Eliminate a DerivStrategy.

mapDerivStrategy :: p ~ GhcPass pass => (XViaStrategy p -> XViaStrategy p) -> DerivStrategy p -> DerivStrategy p

Map over the via type if dealing with ViaStrategy. Otherwise, return the DerivStrategy unchanged.

data XViaStrategyPs

Constructors

XViaStrategyPs (EpAnn [AddEpAnn]) (LHsSigType GhcPs)

RULE declarations

type LRuleDecls pass = XRec pass (RuleDecls pass)

Located Rule Declarations

data RuleDecls pass

Rule Declarations

Constructors

HsRules

Fields

rds_ext :: XCRuleDecls pass

rds_src :: SourceText

rds_rules :: [LRuleDecl pass]

XRuleDecls !(XXRuleDecls pass)

data RuleDecl pass

Rule Declaration

Constructors

HsRule

Fields

rd_ext :: XHsRule pass

After renamer, free-vars from the LHS and RHS

rd_name :: XRec pass (SourceText, RuleName)

Note [Pragma source text] in GHC.Types.Basic

rd_act :: Activation

rd_tyvs :: Maybe [LHsTyVarBndr () (NoGhcTc pass)]

Forall'd type vars

rd_tmvs :: [LRuleBndr pass]

Forall'd term vars, before typechecking; after typechecking this includes all forall'd vars

rd_lhs :: XRec pass (HsExpr pass)

rd_rhs :: XRec pass (HsExpr pass)

AnnKeywordId : AnnOpen,AnnTilde, AnnVal, AnnClose, AnnForall,AnnDot, AnnEqual,

XRuleDecl !(XXRuleDecl pass)

type LRuleDecl pass = XRec pass (RuleDecl pass)

Located Rule Declaration

data HsRuleRn

Constructors

HsRuleRn NameSet NameSet

data HsRuleAnn

Constructors

HsRuleAnn

Fields

ra_tyanns :: Maybe (AddEpAnn, AddEpAnn)

The locations of forall and . for forall'd type vars Using AddEpAnn to capture possible unicode variants

ra_tmanns :: Maybe (AddEpAnn, AddEpAnn)

The locations of forall and . for forall'd term vars Using AddEpAnn to capture possible unicode variants

ra_rest :: [AddEpAnn]

data RuleBndr pass

Rule Binder

Constructors

RuleBndr (XCRuleBndr pass) (LIdP pass)

RuleBndrSig (XRuleBndrSig pass) (LIdP pass) (HsPatSigType pass)

XRuleBndr !(XXRuleBndr pass)

AnnKeywordId : AnnOpen, AnnDcolon,AnnClose

type LRuleBndr pass = XRec pass (RuleBndr pass)

Located Rule Binder

collectRuleBndrSigTys :: [RuleBndr pass] -> [HsPatSigType pass]

flattenRuleDecls :: [LRuleDecls (GhcPass p)] -> [LRuleDecl (GhcPass p)]

pprFullRuleName :: Located (SourceText, RuleName) -> SDoc

default declarations

data DefaultDecl pass

Default Declaration

Constructors

DefaultDecl (XCDefaultDecl pass) [LHsType pass]

AnnKeywordIds : AnnDefault, AnnOpen,AnnClose

XDefaultDecl !(XXDefaultDecl pass)

type LDefaultDecl pass = XRec pass (DefaultDecl pass)

Located Default Declaration

Template haskell declaration splice

data SpliceExplicitFlag

Constructors

ExplicitSplice
= $(f x y)

ImplicitSplice
= f x y, i.e. a naked top level expression

data SpliceDecl p

Splice Declaration

Constructors

SpliceDecl (XSpliceDecl p) (XRec p (HsSplice p)) SpliceExplicitFlag

XSpliceDecl !(XXSpliceDecl p)

type LSpliceDecl pass = XRec pass (SpliceDecl pass)

Located Splice Declaration

Foreign function interface declarations

data ForeignDecl pass

Foreign Declaration

Constructors

ForeignImport

Fields

fd_i_ext :: XForeignImport pass

fd_name :: LIdP pass

fd_sig_ty :: LHsSigType pass

fd_fi :: ForeignImport

ForeignExport

AnnKeywordId : AnnForeign, AnnImport,AnnExport, AnnDcolon

Fields

fd_e_ext :: XForeignExport pass

fd_name :: LIdP pass

fd_sig_ty :: LHsSigType pass

fd_fe :: ForeignExport

XForeignDecl !(XXForeignDecl pass)

type LForeignDecl pass = XRec pass (ForeignDecl pass)

Located Foreign Declaration

data ForeignImport

Constructors

CImport (Located CCallConv) (Located Safety) (Maybe Header) CImportSpec (Located SourceText)

data ForeignExport

Constructors

CExport (Located CExportSpec) (Located SourceText)

data CImportSpec

Constructors

CLabel CLabelString

CFunction CCallTarget

CWrapper

Data-constructor declarations

data ConDecl pass

data T b = forall a. Eq a => MkT a b
  MkT :: forall b a. Eq a => MkT a b

data T b where
     MkT1 :: Int -> T Int

data T = Int MkT Int
       | MkT2

data T a where
     Int MkT Int :: T Int

AnnKeywordIds : AnnOpen, AnnDotdot,AnnCLose, AnnEqual,AnnVbar, AnnDarrow,AnnDarrow, AnnForall,AnnDot

data Constructor Declaration

Constructors

ConDeclGADT

Fields

con_g_ext :: XConDeclGADT pass

con_names :: [LIdP pass]

con_bndrs :: XRec pass (HsOuterSigTyVarBndrs pass)

The outermost type variable binders, be they explicit or implicit. The XRec is used to anchor exact print annotations, AnnForall and AnnDot.

con_mb_cxt :: Maybe (LHsContext pass)

User-written context (if any)

con_g_args :: HsConDeclGADTDetails pass

Arguments; never infix

con_res_ty :: LHsType pass

Result type

con_doc :: Maybe LHsDocString

A possible Haddock comment.

ConDeclH98

Fields

con_ext :: XConDeclH98 pass

con_name :: LIdP pass

con_forall :: Bool

True = explicit user-written forall e.g. data T a = forall b. MkT b (b->a) con_ex_tvs = {b} False => con_ex_tvs is empty

con_ex_tvs :: [LHsTyVarBndr Specificity pass]

Existentials only

con_mb_cxt :: Maybe (LHsContext pass)

User-written context (if any)

con_args :: HsConDeclH98Details pass

Arguments; can be infix

con_doc :: Maybe LHsDocString

A possible Haddock comment.

XConDecl !(XXConDecl pass)

type LConDecl pass

Arguments

= XRec pass (ConDecl pass)
May have AnnKeywordId : AnnSemi when in a GADT constructor list

Located data Constructor Declaration

type HsConDeclH98Details pass = HsConDetails Void (HsScaled pass (LBangType pass)) (XRec pass [LConDeclField pass])

The arguments in a Haskell98-style data constructor.

data HsConDeclGADTDetails pass

The arguments in a GADT constructor. Unlike Haskell98-style constructors, GADT constructors cannot be declared with infix syntax. As a result, we do not use HsConDetails here, as InfixCon would be an unrepresentable state. (There is a notion of infix GADT constructors for the purposes of derived Show instances—see Note [Infix GADT constructors] in GHC.Tc.TyCl—but that is an orthogonal concern.)

Constructors

PrefixConGADT [HsScaled pass (LBangType pass)]

RecConGADT (XRec pass [LConDeclField pass])

hsConDeclTheta :: Maybe (LHsContext (GhcPass p)) -> [LHsType (GhcPass p)]

getConNames :: ConDecl GhcRn -> [LocatedN Name]

getRecConArgs_maybe :: ConDecl GhcRn -> Maybe (LocatedL [LConDeclField GhcRn])

Return Just fields if a data constructor declaration uses record syntax (i.e., RecCon), where fields are the field selectors. Otherwise, return Nothing.

Document comments

data DocDecl

Documentation comment Declaration

Constructors

DocCommentNext HsDocString

DocCommentPrev HsDocString

DocCommentNamed String HsDocString

DocGroup Int HsDocString

type LDocDecl pass = XRec pass DocDecl

Located Documentation comment Declaration

docDeclDoc :: DocDecl -> HsDocString

Deprecations

data WarnDecl pass

Warning pragma Declaration

Constructors

Warning (XWarning pass) [LIdP pass] WarningTxt

XWarnDecl !(XXWarnDecl pass)

type LWarnDecl pass = XRec pass (WarnDecl pass)

Located Warning pragma Declaration

data WarnDecls pass

Warning pragma Declarations

Constructors

Warnings

Fields

wd_ext :: XWarnings pass

wd_src :: SourceText

wd_warnings :: [LWarnDecl pass]

XWarnDecls !(XXWarnDecls pass)

type LWarnDecls pass = XRec pass (WarnDecls pass)

Located Warning Declarations

Annotations

data AnnDecl pass

Annotation Declaration

Constructors

HsAnnotation (XHsAnnotation pass) SourceText (AnnProvenance pass) (XRec pass (HsExpr pass))

AnnKeywordId : AnnOpen, AnnType AnnModule AnnClose

XAnnDecl !(XXAnnDecl pass)

type LAnnDecl pass = XRec pass (AnnDecl pass)

Located Annotation Declaration

data AnnProvenance pass

Annotation Provenance

Constructors

ValueAnnProvenance (LIdP pass)

TypeAnnProvenance (LIdP pass)

ModuleAnnProvenance

annProvenanceName_maybe :: forall p. UnXRec p => AnnProvenance p -> Maybe (IdP p)

Role annotations

data RoleAnnotDecl pass

Role Annotation Declaration

Constructors

RoleAnnotDecl (XCRoleAnnotDecl pass) (LIdP pass) [XRec pass (Maybe Role)]

AnnKeywordId : AnnType, AnnRole

XRoleAnnotDecl !(XXRoleAnnotDecl pass)

type LRoleAnnotDecl pass = XRec pass (RoleAnnotDecl pass)

Located Role Annotation Declaration

roleAnnotDeclName :: RoleAnnotDecl (GhcPass p) -> IdP (GhcPass p)

Injective type families

data FamilyResultSig pass

type Family Result Signature

Constructors

NoSig (XNoSig pass)

AnnKeywordId :

KindSig (XCKindSig pass) (LHsKind pass)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP

TyVarSig (XTyVarSig pass) (LHsTyVarBndr () pass)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP, AnnEqual

XFamilyResultSig !(XXFamilyResultSig pass)

type LFamilyResultSig pass = XRec pass (FamilyResultSig pass)

Located type Family Result Signature

data InjectivityAnn pass

If the user supplied an injectivity annotation it is represented using InjectivityAnn. At the moment this is a single injectivity condition - see Note [Injectivity annotation]. `Located name` stores the LHS of injectivity condition. `[Located name]` stores the RHS of injectivity condition. Example:

type family Foo a b c = r | r -> a c where ...

This will be represented as "InjectivityAnn r [a, c]"

Constructors

InjectivityAnn (XCInjectivityAnn pass) (LIdP pass) [LIdP pass]

AnnKeywordId : AnnRarrow, AnnVbar

XInjectivityAnn !(XXInjectivityAnn pass)

type LInjectivityAnn pass = XRec pass (InjectivityAnn pass)

Located Injectivity Annotation

resultVariableName :: FamilyResultSig (GhcPass a) -> Maybe (IdP (GhcPass a))

Maybe return name of the result type variable

familyDeclLName :: FamilyDecl (GhcPass p) -> XRec (GhcPass p) (IdP (GhcPass p))

familyDeclName :: FamilyDecl (GhcPass p) -> IdP (GhcPass p)

Grouping

data HsGroup p

Haskell Group

A HsDecl is categorised into a HsGroup before being fed to the renamer.

Constructors

HsGroup

Fields

hs_ext :: XCHsGroup p

hs_valds :: HsValBinds p

hs_splcds :: [LSpliceDecl p]

hs_tyclds :: [TyClGroup p]

hs_derivds :: [LDerivDecl p]

hs_fixds :: [LFixitySig p]

hs_defds :: [LDefaultDecl p]

hs_fords :: [LForeignDecl p]

hs_warnds :: [LWarnDecls p]

hs_annds :: [LAnnDecl p]

hs_ruleds :: [LRuleDecls p]

hs_docs :: [LDocDecl p]

XHsGroup !(XXHsGroup p)

emptyRdrGroup :: HsGroup (GhcPass p)

emptyRnGroup :: HsGroup (GhcPass p)

appendGroups :: HsGroup (GhcPass p) -> HsGroup (GhcPass p) -> HsGroup (GhcPass p)

hsGroupInstDecls :: HsGroup id -> [LInstDecl id]

hsGroupTopLevelFixitySigs :: HsGroup (GhcPass p) -> [LFixitySig (GhcPass p)]

The fixity signatures for each top-level declaration and class method in an HsGroup. See Note [Top-level fixity signatures in an HsGroup]

partitionBindsAndSigs :: [LHsDecl GhcPs] -> (LHsBinds GhcPs, [LSig GhcPs], [LFamilyDecl GhcPs], [LTyFamInstDecl GhcPs], [LDataFamInstDecl GhcPs], [LDocDecl GhcPs])

Partition a list of HsDecls into function/pattern bindings, signatures, type family declarations, type family instances, and documentation comments.

Panics when given a declaration that cannot be put into any of the output groups.

The primary use of this function is to implement cvBindsAndSigs.

Orphan instances

OutputableBndrId p => Outputable (AnnDecl (GhcPass p))

Instance details

Methods

ppr :: AnnDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (ClsInstDecl (GhcPass p))

Instance details

Methods

ppr :: ClsInstDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (ConDecl (GhcPass p))

Instance details

Methods

ppr :: ConDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (DataFamInstDecl (GhcPass p))

Instance details

Methods

ppr :: DataFamInstDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (DefaultDecl (GhcPass p))

Instance details

Methods

ppr :: DefaultDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (DerivClauseTys (GhcPass p))

Instance details

Methods

ppr :: DerivClauseTys (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (DerivDecl (GhcPass p))

Instance details

Methods

ppr :: DerivDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (DerivStrategy (GhcPass p))

Instance details

Methods

ppr :: DerivStrategy (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (FamilyDecl (GhcPass p))

Instance details

Methods

ppr :: FamilyDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (ForeignDecl (GhcPass p))

Instance details

Methods

ppr :: ForeignDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (FunDep (GhcPass p))

Instance details

Methods

ppr :: FunDep (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsDataDefn (GhcPass p))

Instance details

Methods

ppr :: HsDataDefn (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsDecl (GhcPass p))

Instance details

Methods

ppr :: HsDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsDerivingClause (GhcPass p))

Instance details

Methods

ppr :: HsDerivingClause (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsGroup (GhcPass p))

Instance details

Methods

ppr :: HsGroup (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (InstDecl (GhcPass p))

Instance details

Methods

ppr :: InstDecl (GhcPass p) -> SDoc

OutputableBndr (IdP (GhcPass p)) => Outputable (RoleAnnotDecl (GhcPass p))

Instance details

Methods

ppr :: RoleAnnotDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (RuleBndr (GhcPass p))

Instance details

Methods

ppr :: RuleBndr (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (RuleDecl (GhcPass p))

Instance details

Methods

ppr :: RuleDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (RuleDecls (GhcPass p))

Instance details

Methods

ppr :: RuleDecls (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (SpliceDecl (GhcPass p))

Instance details

Methods

ppr :: SpliceDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (StandaloneKindSig (GhcPass p))

Instance details

Methods

ppr :: StandaloneKindSig (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (TyClDecl (GhcPass p))

Instance details

Methods

ppr :: TyClDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (TyClGroup (GhcPass p))

Instance details

Methods

ppr :: TyClGroup (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (TyFamInstDecl (GhcPass p))

Instance details

Methods

ppr :: TyFamInstDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (WarnDecl (GhcPass p))

Instance details

Methods

ppr :: WarnDecl (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (WarnDecls (GhcPass p))

Instance details

Methods

ppr :: WarnDecls (GhcPass p) -> SDoc