GHC.Hs.Doc

Documentation

data HsDocString

Haskell Documentation String

Internally this is a UTF8-Encoded ByteString.

type LHsDocString = Located HsDocString

Located Haskell Documentation String

mkHsDocString :: String -> HsDocString

mkHsDocStringUtf8ByteString :: ByteString -> HsDocString

Create a HsDocString from a UTF8-encoded ByteString.

isEmptyDocString :: HsDocString -> Bool

unpackHDS :: HsDocString -> String

hsDocStringToByteString :: HsDocString -> ByteString

Return the contents of a HsDocString as a UTF8-encoded ByteString.

ppr_mbDoc :: Maybe LHsDocString -> SDoc

appendDocs :: HsDocString -> HsDocString -> HsDocString

Join two docstrings.

Non-empty docstrings are joined with two newlines in between, resulting in separate paragraphs.

concatDocs :: [HsDocString] -> Maybe HsDocString

Concat docstrings with two newlines in between.

Empty docstrings are skipped.

If all inputs are empty, Nothing is returned.

newtype DeclDocMap

Docs for declarations: functions, data types, instances, methods etc.

Constructors

DeclDocMap (Map Name HsDocString)

emptyDeclDocMap :: DeclDocMap

newtype ArgDocMap

Docs for arguments. E.g. function arguments, method arguments.

Constructors

ArgDocMap (Map Name (IntMap HsDocString))

emptyArgDocMap :: ArgDocMap

data ExtractedTHDocs

Maps of docs that were added via Template Haskell's putDoc.

Constructors

ExtractedTHDocs

Fields

ethd_mod_header :: Maybe HsDocString

The added module header documentation, if it exists.

ethd_decl_docs :: DeclDocMap

The documentation added to declarations.

ethd_arg_docs :: ArgDocMap

The documentation added to function arguments.

ethd_inst_docs :: DeclDocMap

The documentation added to class and family instances.