GHC.Hs.Expr

Description

Abstract Haskell syntax for expressions.

Documentation

module Language.Haskell.Syntax.Expr

arrowToHsExpr :: HsArrowOf (LocatedA (HsExpr GhcRn)) GhcRn -> LocatedA (HsExpr GhcRn)

gHsPar :: forall (p :: Pass). IsPass p => LHsExpr (GhcPass p) -> HsExpr (GhcPass p)

Parenthesize an expression without token information

hsExprNeedsParens :: forall (p :: Pass). IsPass p => PprPrec -> HsExpr (GhcPass p) -> Bool

hsExprNeedsParens p e returns True if the expression e needs parentheses under precedence p.

hsLMatchPats :: forall (id :: Pass) body. LMatch (GhcPass id) body -> [LPat (GhcPass id)]

isAtomicHsExpr :: forall (p :: Pass). IsPass p => HsExpr (GhcPass p) -> Bool

isEmptyMatchGroup :: forall (p :: Pass) body. MatchGroup (GhcPass p) body -> Bool

isHsThingRnExpr :: HsThingRn -> Bool

isHsThingRnPat :: HsThingRn -> Bool

isHsThingRnStmt :: HsThingRn -> Bool

isQuietHsCmd :: HsCmd id -> Bool

isQuietHsExpr :: HsExpr id -> Bool

isSingletonMatchGroup :: forall (p :: Pass) body. [LMatch (GhcPass p) body] -> Bool

Is there only one RHS in this list of matches?

lamCaseKeyword :: HsLamVariant -> SDoc

matchArrowContextErrString :: HsArrowMatchContext -> SDoc

matchContextErrString :: Outputable fn => HsMatchContext fn -> SDoc

matchDoContextErrString :: HsDoFlavour -> SDoc

matchGroupArity :: forall (id :: Pass) body. MatchGroup (GhcPass id) body -> Arity

matchSeparator :: HsMatchContext fn -> SDoc

mkExpandedExpr

Arguments

:: HsExpr GhcRn
source expression

-> HsExpr GhcRn
expanded expression

-> HsExpr GhcRn
suitably wrapped XXExprGhcRn

Build an expression using the extension constructor XExpr, and the two components of the expansion: original expression and expanded expressions.

mkExpandedExprTc

Arguments

:: HsExpr GhcRn
source expression

-> HsExpr GhcTc
expanded typechecked expression

-> HsExpr GhcTc
suitably wrapped XXExprGhcRn

Build a XXExprGhcRn out of an extension constructor, and the two components of the expansion: original and expanded typechecked expressions.

mkExpandedPatRn

Arguments

:: LPat GhcRn
source pattern

-> HsExpr GhcRn
expanded expression

-> HsExpr GhcRn
suitably wrapped XXExprGhcRn

mkExpandedStmt

Arguments

:: ExprLStmt GhcRn
source statement

-> HsExpr GhcRn
expanded expression

-> HsExpr GhcRn
suitably wrapped XXExprGhcRn

Build an expression using the extension constructor XExpr, and the two components of the expansion: original do stmt and expanded expression

mkExpandedStmtAt

Arguments

:: SrcSpanAnnA
Location for the expansion expression

-> ExprLStmt GhcRn
source statement

-> HsExpr GhcRn
expanded expression

-> LHsExpr GhcRn
suitably wrapped located XXExprGhcRn

Build an expression using the extension constructor XExpr, and the two components of the expansion: original do stmt and expanded expression an associate with a provided location

mkExpandedStmtPopAt

Arguments

:: SrcSpanAnnA
Location for the expansion statement

-> ExprLStmt GhcRn
source statement

-> HsExpr GhcRn
expanded expression

-> LHsExpr GhcRn
suitably wrapped XXExprGhcRn

Wrap the expanded version of the expression with a pop.

mkExpandedStmtTc

Arguments

:: ExprLStmt GhcRn
source do statement

-> HsExpr GhcTc
expanded typechecked expression

-> HsExpr GhcTc
suitably wrapped XXExprGhcRn

Build a XXExprGhcRn out of an extension constructor. The two components of the expansion are: original statement and expanded typechecked expression.

mkPopErrCtxtExpr :: LHsExpr GhcRn -> HsExpr GhcRn

Wrap a located expression with a PopErrCtxt

mkPopErrCtxtExprAt :: SrcSpanAnnA -> LHsExpr GhcRn -> LHsExpr GhcRn

Wrap a located expression with a PopSrcExpr with an appropriate location

mkRnSyntaxExpr :: Name -> SyntaxExprRn

Make a SyntaxExpr from a Name (the "rn" is because this is used in the renamer).

mkSyntaxExpr :: HsExpr GhcRn -> SyntaxExprRn

Make a 'SyntaxExpr GhcRn' from an expression Used only in getMonadFailOp. See Note [Monad fail : Rebindable syntax, overloaded strings] in GHC.Rename.Expr

noExpr :: forall (p :: Pass). HsExpr (GhcPass p)

This is used for rebindable-syntax pieces that are too polymorphic for tcSyntaxOp (trS_fmap and the mzip in ParStmt)

noSyntaxExpr :: forall (p :: Pass). IsPass p => SyntaxExpr (GhcPass p)

parenthesizeHsExpr :: forall (p :: Pass). IsPass p => PprPrec -> LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

parenthesizeHsExpr p e checks if hsExprNeedsParens p e is true, and if so, surrounds e with an HsPar. Otherwise, it simply returns e.

pp_dotdot :: SDoc

pp_rhs :: Outputable body => HsMatchContext fn -> body -> SDoc

pprAHsDoFlavour :: HsDoFlavour -> SDoc

pprAStmtContext :: Outputable fn => HsStmtContext fn -> SDoc

pprArg :: forall (idL :: Pass). OutputableBndrId idL => ApplicativeArg (GhcPass idL) -> SDoc

pprArrowExpr :: forall (p :: Pass) body. (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => [LStmt (GhcPass p) body] -> SDoc

pprArrowMatchContextNoun :: HsArrowMatchContext -> SDoc

pprArrowMatchContextNouns :: HsArrowMatchContext -> SDoc

pprBindStmt :: (Outputable pat, Outputable expr) => pat -> expr -> SDoc

pprBinds :: forall (idL :: Pass) (idR :: Pass). (OutputableBndrId idL, OutputableBndrId idR) => HsLocalBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

pprBy :: Outputable body => Maybe body -> SDoc

pprCmd :: forall (p :: Pass). OutputableBndrId p => HsCmd (GhcPass p) -> SDoc

pprCmdArg :: forall (p :: Pass). OutputableBndrId p => HsCmdTop (GhcPass p) -> SDoc

pprComp :: forall (p :: Pass) body. (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => [LStmt (GhcPass p) body] -> SDoc

pprDebugParendExpr :: forall (p :: Pass). OutputableBndrId p => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprDo :: forall (p :: Pass) body. (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => HsDoFlavour -> [LStmt (GhcPass p) body] -> SDoc

pprExpr :: forall (p :: Pass). OutputableBndrId p => HsExpr (GhcPass p) -> SDoc

pprExternalSrcLoc :: (StringLiteral, (Int, Int), (Int, Int)) -> SDoc

pprFieldLabelStrings :: (UnXRec p, Outputable (XRec p FieldLabelString)) => FieldLabelStrings p -> SDoc

pprFunBind :: forall (idR :: Pass). OutputableBndrId idR => MatchGroup (GhcPass idR) (LHsExpr (GhcPass idR)) -> SDoc

pprGRHS :: forall (idR :: Pass) body fn. (OutputableBndrId idR, Outputable body) => HsMatchContext fn -> GRHS (GhcPass idR) body -> SDoc

pprGRHSs :: forall (idR :: Pass) body fn. (OutputableBndrId idR, Outputable body) => HsMatchContext fn -> GRHSs (GhcPass idR) body -> SDoc

pprHsArrType :: HsArrAppType -> SDoc

pprHsDoFlavour :: HsDoFlavour -> SDoc

pprLCmd :: forall (p :: Pass). OutputableBndrId p => LHsCmd (GhcPass p) -> SDoc

pprLExpr :: forall (p :: Pass). OutputableBndrId p => LHsExpr (GhcPass p) -> SDoc

pprMatch :: forall (idR :: Pass) body. (OutputableBndrId idR, Outputable body) => Match (GhcPass idR) body -> SDoc

pprMatchContext :: Outputable fn => HsMatchContext fn -> SDoc

pprMatchContextNoun :: Outputable fn => HsMatchContext fn -> SDoc

pprMatchContextNouns :: Outputable fn => HsMatchContext fn -> SDoc

pprMatchInCtxt :: forall (idR :: Pass) body. (OutputableBndrId idR, Outputable body) => Match (GhcPass idR) body -> SDoc

pprMatches :: forall (idR :: Pass) body. (OutputableBndrId idR, Outputable body) => MatchGroup (GhcPass idR) body -> SDoc

pprParendExpr :: forall (p :: Pass). OutputableBndrId p => PprPrec -> HsExpr (GhcPass p) -> SDoc

pprParendLExpr :: forall (p :: Pass). OutputableBndrId p => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprPatBind :: forall (bndr :: Pass) (p :: Pass). (OutputableBndrId bndr, OutputableBndrId p) => LPat (GhcPass bndr) -> GRHSs (GhcPass p) (LHsExpr (GhcPass p)) -> SDoc

pprPendingSplice :: forall (p :: Pass). OutputableBndrId p => SplicePointName -> LHsExpr (GhcPass p) -> SDoc

pprPrefixFastString :: FastString -> SDoc

pprQuals :: forall (p :: Pass) body. (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => [LStmt (GhcPass p) body] -> SDoc

pprStmt :: forall (idL :: Pass) (idR :: Pass) body. (OutputableBndrId idL, OutputableBndrId idR, Anno (StmtLR (GhcPass idL) (GhcPass idR) body) ~ SrcSpanAnnA, Outputable body) => StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc

pprStmtCat :: forall (p :: Pass) body. IsPass p => Stmt (GhcPass p) body -> SDoc

pprStmtContext :: Outputable fn => HsStmtContext fn -> SDoc

pprStmtInCtxt :: forall (idL :: Pass) (idR :: Pass) fn body. (OutputableBndrId idL, OutputableBndrId idR, Outputable fn, Outputable body, Anno (StmtLR (GhcPass idL) (GhcPass idR) body) ~ SrcSpanAnnA) => HsStmtContext fn -> StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc

pprTransStmt :: Outputable body => Maybe body -> body -> TransForm -> SDoc

pprTransformStmt :: forall (p :: Pass). OutputableBndrId p => [IdP (GhcPass p)] -> LHsExpr (GhcPass p) -> Maybe (LHsExpr (GhcPass p)) -> SDoc

pprTypedSplice :: forall (p :: Pass). OutputableBndrId p => Maybe SplicePointName -> LHsExpr (GhcPass p) -> SDoc

pprUntypedSplice :: forall (p :: Pass). OutputableBndrId p => Bool -> Maybe SplicePointName -> HsUntypedSplice (GhcPass p) -> SDoc

ppr_apps :: forall (p :: Pass). OutputableBndrId p => HsExpr (GhcPass p) -> [Either (LHsExpr (GhcPass p)) (LHsWcType (NoGhcTc (GhcPass p)))] -> SDoc

ppr_cmd :: forall (p :: Pass). OutputableBndrId p => HsCmd (GhcPass p) -> SDoc

ppr_do_stmts :: forall (idL :: Pass) (idR :: Pass) body. (OutputableBndrId idL, OutputableBndrId idR, Anno (StmtLR (GhcPass idL) (GhcPass idR) body) ~ SrcSpanAnnA, Outputable body) => [LStmtLR (GhcPass idL) (GhcPass idR) body] -> SDoc

ppr_expr :: forall (p :: Pass). OutputableBndrId p => HsExpr (GhcPass p) -> SDoc

ppr_infix_expr :: forall (p :: Pass). OutputableBndrId p => HsExpr (GhcPass p) -> Maybe SDoc

ppr_infix_expr_rn :: XXExprGhcRn -> Maybe SDoc

ppr_infix_expr_tc :: XXExprGhcTc -> Maybe SDoc

ppr_infix_hs_expansion :: HsThingRn -> Maybe SDoc

ppr_lcmd :: forall (p :: Pass). OutputableBndrId p => LHsCmd (GhcPass p) -> SDoc

ppr_lexpr :: forall (p :: Pass). OutputableBndrId p => LHsExpr (GhcPass p) -> SDoc

ppr_module_name_prefix :: Maybe ModuleName -> SDoc

ppr_quasi :: OutputableBndr p => p -> FastString -> SDoc

ppr_splice :: forall (p :: Pass). OutputableBndrId p => SDoc -> Maybe SplicePointName -> LHsExpr (GhcPass p) -> SDoc

ppr_with_pending_tc_splices :: SDoc -> [PendingTcSplice] -> SDoc

prependQualified :: Maybe ModuleName -> SDoc -> SDoc

stripParensHsExpr :: forall (p :: Pass). HsExpr (GhcPass p) -> HsExpr (GhcPass p)

stripParensLHsExpr :: forall (p :: Pass). LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

thBrackets :: SDoc -> SDoc -> SDoc

thTyBrackets :: SDoc -> SDoc

tupArgPresent :: forall (p :: Pass). HsTupArg (GhcPass p) -> Bool

tupArgPresent_maybe :: forall (p :: Pass). HsTupArg (GhcPass p) -> Maybe (LHsExpr (GhcPass p))

tupArgsPresent_maybe :: forall (p :: Pass). [HsTupArg (GhcPass p)] -> Maybe [LHsExpr (GhcPass p)]

data AnnArithSeq

Constructors

AnnArithSeq

Fields

aas_open :: EpToken "["

aas_comma :: Maybe (EpToken ",")

aas_dotdot :: EpToken ".."

aas_close :: EpToken "]"

data AnnExplicitSum

Constructors

AnnExplicitSum

Fields

aesOpen :: EpaLocation

aesBarsBefore :: [EpToken "|"]

aesBarsAfter :: [EpToken "|"]

aesClose :: EpaLocation

data AnnFieldLabel

Constructors

AnnFieldLabel

Fields

afDot :: Maybe (EpToken ".")

data AnnFunRhs

Constructors

AnnFunRhs

Fields

afr_strict :: EpToken "!"

afr_opens :: [EpToken "("]

afr_closes :: [EpToken ")"]

data AnnProjection

Constructors

AnnProjection

Fields

apOpen :: EpToken "("

apClose :: EpToken ")"

data AnnTransStmt

Constructors

AnnTransStmt

Fields

ats_then :: EpToken "then"

ats_group :: Maybe (EpToken "group")

ats_by :: Maybe (EpToken "by")

ats_using :: Maybe (EpToken "using")

data AnnsIf

Constructors

AnnsIf

Fields

aiIf :: EpToken "if"

aiThen :: EpToken "then"

aiElse :: EpToken "else"

aiThenSemi :: Maybe (EpToken ";")

aiElseSemi :: Maybe (EpToken ";")

data ApplicativeArg idL

Applicative Argument

Constructors

ApplicativeArgOne

Fields

xarg_app_arg_one :: XApplicativeArgOne idL

The fail operator, after renaming

The fail operator is needed if this is a BindStmt where the pattern can fail. E.g.: (Just a) <- stmt The fail operator will be invoked if the pattern match fails. It is also used for guards in MonadComprehensions. The fail operator is Nothing if the pattern match can't fail

app_arg_pattern :: LPat idL

arg_expr :: LHsExpr idL

is_body_stmt :: Bool

True = was a BodyStmt, False = was a BindStmt. See Note [Applicative BodyStmt]

ApplicativeArgMany

Fields

xarg_app_arg_many :: XApplicativeArgMany idL

app_stmts :: [ExprLStmt idL]

final_expr :: HsExpr idL

bv_pattern :: LPat idL

stmt_context :: HsDoFlavour

context of the do expression, used in pprArg

XApplicativeArg !(XXApplicativeArg idL)

data ApplicativeStmt idL idR

ApplicativeStmt represents an applicative expression built with <$> and <*>. It is generated by the renamer, and is desugared into the appropriate applicative expression by the desugarer, but it is intended to be invisible in error messages.

For full details, see Note [ApplicativeDo] in GHC.Rename.Expr

Constructors

ApplicativeStmt (XApplicativeStmt idL idR) [(SyntaxExpr idR, ApplicativeArg idL)] (Maybe (SyntaxExpr idR))

data BracketAnn noE hasE

Constructors

BracketNoE noE

BracketHasE hasE

type CmdSyntaxTable p = [(Name, HsExpr p)]

Command Syntax Table (for Arrow syntax)

data CmdTopTc

Constructors

CmdTopTc Type Type (CmdSyntaxTable GhcTc)

data DelayedSplice

Constructors

DelayedSplice TcLclEnv (LHsExpr GhcRn) TcType (LHsExpr GhcTc)

data EpAnnHsCase

Constructors

EpAnnHsCase

Fields

hsCaseAnnCase :: EpToken "case"

hsCaseAnnOf :: EpToken "of"

data EpAnnLam

Constructors

EpAnnLam

Fields

epl_lambda :: EpToken "\\"

Location of '\' keyword

epl_case :: Maybe EpaLocation

Location of 'case' or cases keyword, depending on related HsLamVariant.

data EpAnnUnboundVar

Constructors

EpAnnUnboundVar

Fields

hsUnboundBackquotes :: (EpToken "`", EpToken "`")

hsUnboundHole :: EpToken "_"

data GrhsAnn

Constructors

GrhsAnn

Fields

ga_vbar :: Maybe (EpToken "|")

ga_sep :: Either (EpToken "=") TokRarrow

Match separator location, `=` or ->

data HsBracketTc

Constructors

HsBracketTc

Fields

hsb_quote :: HsQuote GhcRn

hsb_ty :: Type

hsb_wrap :: Maybe QuoteWrapper

hsb_splices :: [PendingTcSplice]

type HsMatchContextPs = HsMatchContext (LIdP GhcPs)

type HsMatchContextRn = HsMatchContext (LIdP GhcRn)

data family HsRecUpdParent x

Information about the parent of a record update:

the parent type constructor or pattern synonym,

the relevant con-likes,

the field labels.

type HsStmtContextRn = HsStmtContext (LIdP GhcRn)

data HsThingRn

The different source constructs that we use to instantiate the "original" field in an `XXExprGhcRn original expansion`

Constructors

OrigExpr (HsExpr GhcRn)

OrigStmt (ExprLStmt GhcRn)

OrigPat (LPat GhcRn)

data HsUntypedSpliceResult thing

Constructors

HsUntypedSpliceTop

Fields

utsplice_result_finalizers :: ThModFinalizers

TH finalizers produced by the splice.

utsplice_result :: thing

The result of splicing; See Note [Lifecycle of a splice]

HsUntypedSpliceNested SplicePointName

data HsWrap (hs_syn :: Type -> Type)

HsWrap appears only in typechecker output

Constructors

HsWrap HsWrapper (hs_syn GhcTc)

data MatchGroupTc

Constructors

MatchGroupTc

Fields

mg_arg_tys :: [Scaled Type]

mg_res_ty :: Type

mg_origin :: Origin

data PendingRnSplice

Pending Renamer Splice

Constructors

PendingRnSplice UntypedSpliceFlavour SplicePointName (LHsExpr GhcRn)

data PendingTcSplice

Pending Type-checker Splice

Constructors

PendingTcSplice SplicePointName (LHsExpr GhcTc)

type PostTcExpr = HsExpr GhcTc

Post-Type checking Expression

PostTcExpr is an evidence expression attached to the syntax tree by the type checker (c.f. postTcType).

type PostTcTable = [(Name, PostTcExpr)]

Post-Type checking Table

We use a PostTcTable where there are a bunch of pieces of evidence, more than is convenient to keep individually.

data RecStmtTc

Constructors

RecStmtTc

Fields

recS_bind_ty :: Type

recS_later_rets :: [PostTcExpr]

recS_rec_rets :: [PostTcExpr]

recS_ret_ty :: Type

type SplicePointName = Name

type family SyntaxExprGhc (p :: Pass) = (r :: Type) | r -> p where ...

Equations

SyntaxExprGhc 'Parsed = NoExtField

SyntaxExprGhc 'Renamed = SyntaxExprRn

SyntaxExprGhc 'Typechecked = SyntaxExprTc

data SyntaxExprRn

The function to use in rebindable syntax. See Note [NoSyntaxExpr].

Constructors

SyntaxExprRn (HsExpr GhcRn)

NoSyntaxExprRn

data SyntaxExprTc

An expression with wrappers, used for rebindable syntax

This should desugar to

syn_res_wrap $ syn_expr (syn_arg_wraps[0] arg0)
                        (syn_arg_wraps[1] arg1) ...

where the actual arguments come from elsewhere in the AST.

Constructors

SyntaxExprTc

Fields

syn_expr :: HsExpr GhcTc

syn_arg_wraps :: [HsWrapper]

syn_res_wrap :: HsWrapper

NoSyntaxExprTc

newtype ThModFinalizers

Finalizers produced by a splice with addModFinalizer

See Note [Delaying modFinalizers in untyped splices] in GHC.Rename.Splice. For how this is used.

Constructors

ThModFinalizers [ForeignRef (Q ())]

data UntypedSpliceFlavour

Constructors

UntypedExpSplice

UntypedPatSplice

UntypedTypeSplice

UntypedDeclSplice

type family XApplicativeArgMany x

type family XApplicativeArgOne x

type family XApplicativeStmt x x'

data XBindStmtRn

Constructors

XBindStmtRn

Fields

xbsrn_bindOp :: SyntaxExpr GhcRn

xbsrn_failOp :: FailOperator GhcRn

data XBindStmtTc

Constructors

XBindStmtTc

Fields

xbstc_bindOp :: SyntaxExpr GhcTc

xbstc_boundResultType :: Type

xbstc_boundResultMult :: Mult

xbstc_failOp :: FailOperator GhcTc

type family XXApplicativeArg x

data XXExprGhcRn

Constructors

ExpandedThingRn

Fields

xrn_orig :: HsThingRn

xrn_expanded :: HsExpr GhcRn

PopErrCtxt !(LHsExpr GhcRn)

HsRecSelRn (FieldOcc GhcRn)
Variable pointing to record selector See Note [Non-overloaded record field selectors] and Note [Record selectors in the AST]

data XXExprGhcTc

Constructors

WrapExpr HsWrapper (HsExpr GhcTc)

ExpandedThingTc

Fields

xtc_orig :: HsThingRn

xtc_expanded :: HsExpr GhcTc

ConLikeTc ConLike [TcTyVar] [Scaled TcType]

HsTick CoreTickish (LHsExpr GhcTc)

HsBinTick Int Int (LHsExpr GhcTc)

HsRecSelTc (FieldOcc GhcTc)
Variable pointing to record selector See Note [Non-overloaded record field selectors] and Note [Record selectors in the AST]

Orphan instances

Outputable HsArrowMatchContext

Instance details

Methods

ppr :: HsArrowMatchContext -> SDoc

Outputable HsLamVariant

Instance details

Methods

ppr :: HsLamVariant -> SDoc

OutputableBndrId p => Outputable (ArithSeqInfo (GhcPass p))

Instance details

Methods

ppr :: ArithSeqInfo (GhcPass p) -> SDoc

UnXRec p => Outputable (DotFieldOcc p)

Instance details

Methods

ppr :: DotFieldOcc p -> SDoc

(UnXRec p, Outputable (XRec p FieldLabelString)) => Outputable (FieldLabelStrings p)

Instance details

Methods

ppr :: FieldLabelStrings p -> SDoc

OutputableBndrId p => Outputable (HsCmd (GhcPass p))

Instance details

Methods

ppr :: HsCmd (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsCmdTop (GhcPass p))

Instance details

Methods

ppr :: HsCmdTop (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsExpr (GhcPass p))

Instance details

Methods

ppr :: HsExpr (GhcPass p) -> SDoc

Outputable fn => Outputable (HsMatchContext fn)

Instance details

Methods

ppr :: HsMatchContext fn -> SDoc

Outputable (HsPragE (GhcPass p))

Instance details

Methods

ppr :: HsPragE (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsQuote (GhcPass p))

Instance details

Methods

ppr :: HsQuote (GhcPass p) -> SDoc

Outputable fn => Outputable (HsStmtContext fn)

Instance details

Methods

ppr :: HsStmtContext fn -> SDoc

(UnXRec p, Outputable (XRec p FieldLabelString)) => OutputableBndr (Located (FieldLabelStrings p))

Instance details

Methods

pprBndr :: BindingSite -> Located (FieldLabelStrings p) -> SDoc

pprPrefixOcc :: Located (FieldLabelStrings p) -> SDoc

pprInfixOcc :: Located (FieldLabelStrings p) -> SDoc

bndrIsJoin_maybe :: Located (FieldLabelStrings p) -> JoinPointHood

(UnXRec p, Outputable (XRec p FieldLabelString)) => OutputableBndr (FieldLabelStrings p)

Instance details

Methods

pprBndr :: BindingSite -> FieldLabelStrings p -> SDoc

pprPrefixOcc :: FieldLabelStrings p -> SDoc

pprInfixOcc :: FieldLabelStrings p -> SDoc

bndrIsJoin_maybe :: FieldLabelStrings p -> JoinPointHood

HasAnnotation (Anno a) => WrapXRec (GhcPass p) a

Instance details

Methods

wrapXRec :: a -> XRec (GhcPass p) a

(OutputableBndrId pr, Outputable body) => Outputable (Match (GhcPass pr) body)

Instance details

Methods

ppr :: Match (GhcPass pr) body -> SDoc

(Outputable (StmtLR (GhcPass idL) (GhcPass idL) (LHsExpr (GhcPass idL))), Outputable (XXParStmtBlock (GhcPass idL) (GhcPass idR))) => Outputable (ParStmtBlock (GhcPass idL) (GhcPass idR))

Instance details

Methods

ppr :: ParStmtBlock (GhcPass idL) (GhcPass idR) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr, Anno (StmtLR (GhcPass pl) (GhcPass pr) body) ~ SrcSpanAnnA, Outputable body) => Outputable (StmtLR (GhcPass pl) (GhcPass pr) body)

Instance details

Methods

ppr :: StmtLR (GhcPass pl) (GhcPass pr) body -> SDoc