GHC.Hs.Expr

Description

Abstract Haskell syntax for expressions.

Documentation

module Language.Haskell.Syntax.Expr

pprExpr :: OutputableBndrId p => HsExpr (GhcPass p) -> SDoc

pprFunBind :: OutputableBndrId idR => MatchGroup (GhcPass idR) (LHsExpr (GhcPass idR)) -> SDoc

pprLExpr :: OutputableBndrId p => LHsExpr (GhcPass p) -> SDoc

pprPatBind :: forall bndr p. (OutputableBndrId bndr, OutputableBndrId p) => LPat (GhcPass bndr) -> GRHSs (GhcPass p) (LHsExpr (GhcPass p)) -> SDoc

pprSplice :: forall p. OutputableBndrId p => HsSplice (GhcPass p) -> SDoc

pprSpliceDecl :: OutputableBndrId p => HsSplice (GhcPass p) -> SpliceExplicitFlag -> SDoc

hsExprNeedsParens :: forall p. IsPass p => PprPrec -> HsExpr (GhcPass p) -> Bool

hsExprNeedsParens p e returns True if the expression e needs parentheses under precedence p.

hsLMatchPats :: LMatch (GhcPass id) body -> [LPat (GhcPass id)]

isAtomicHsExpr :: forall p. IsPass p => HsExpr (GhcPass p) -> Bool

isEmptyMatchGroup :: MatchGroup (GhcPass p) body -> Bool

isQuietHsCmd :: HsCmd id -> Bool

isQuietHsExpr :: HsExpr id -> Bool

isSingletonMatchGroup :: [LMatch (GhcPass p) body] -> Bool

Is there only one RHS in this list of matches?

matchArrowContextErrString :: HsArrowMatchContext -> SDoc

matchContextErrString :: OutputableBndrId p => HsMatchContext (GhcPass p) -> SDoc

matchGroupArity :: MatchGroup (GhcPass id) body -> Arity

mkRnSyntaxExpr :: Name -> SyntaxExprRn

Make a SyntaxExpr from a Name (the "rn" is because this is used in the renamer).

mkSyntaxExpr :: HsExpr GhcRn -> SyntaxExprRn

Make a 'SyntaxExpr GhcRn' from an expression Used only in getMonadFailOp. See Note [Monad fail : Rebindable syntax, overloaded strings] in GHC.Rename.Expr

noExpr :: HsExpr (GhcPass p)

This is used for rebindable-syntax pieces that are too polymorphic for tcSyntaxOp (trS_fmap and the mzip in ParStmt)

noSyntaxExpr :: forall p. IsPass p => SyntaxExpr (GhcPass p)

parenthesizeHsExpr :: IsPass p => PprPrec -> LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

parenthesizeHsExpr p e checks if hsExprNeedsParens p e is true, and if so, surrounds e with an HsPar. Otherwise, it simply returns e.

pp_dotdot :: SDoc

pp_rhs :: Outputable body => HsMatchContext passL -> body -> SDoc

pprArg :: forall idL. OutputableBndrId idL => ApplicativeArg (GhcPass idL) -> SDoc

pprBindStmt :: (Outputable pat, Outputable expr) => pat -> expr -> SDoc

pprBinds :: (OutputableBndrId idL, OutputableBndrId idR) => HsLocalBindsLR (GhcPass idL) (GhcPass idR) -> SDoc

pprBy :: Outputable body => Maybe body -> SDoc

pprCmd :: OutputableBndrId p => HsCmd (GhcPass p) -> SDoc

pprCmdArg :: OutputableBndrId p => HsCmdTop (GhcPass p) -> SDoc

pprComp :: (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => [LStmt (GhcPass p) body] -> SDoc

pprDebugParendExpr :: OutputableBndrId p => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprDo :: (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => HsStmtContext any -> [LStmt (GhcPass p) body] -> SDoc

pprGRHS :: (OutputableBndrId idR, Outputable body) => HsMatchContext passL -> GRHS (GhcPass idR) body -> SDoc

pprGRHSs :: (OutputableBndrId idR, Outputable body) => HsMatchContext passL -> GRHSs (GhcPass idR) body -> SDoc

pprHsBracket :: OutputableBndrId p => HsBracket (GhcPass p) -> SDoc

pprLCmd :: OutputableBndrId p => LHsCmd (GhcPass p) -> SDoc

pprMatch :: (OutputableBndrId idR, Outputable body) => Match (GhcPass idR) body -> SDoc

pprMatchInCtxt :: (OutputableBndrId idR, Outputable body) => Match (GhcPass idR) body -> SDoc

pprMatches :: (OutputableBndrId idR, Outputable body) => MatchGroup (GhcPass idR) body -> SDoc

pprParendExpr :: OutputableBndrId p => PprPrec -> HsExpr (GhcPass p) -> SDoc

pprParendLExpr :: OutputableBndrId p => PprPrec -> LHsExpr (GhcPass p) -> SDoc

pprPendingSplice :: OutputableBndrId p => SplicePointName -> LHsExpr (GhcPass p) -> SDoc

pprQuals :: (OutputableBndrId p, Outputable body, Anno (StmtLR (GhcPass p) (GhcPass p) body) ~ SrcSpanAnnA) => [LStmt (GhcPass p) body] -> SDoc

pprStmt :: forall idL idR body. (OutputableBndrId idL, OutputableBndrId idR, Anno (StmtLR (GhcPass idL) (GhcPass idR) body) ~ SrcSpanAnnA, Outputable body) => StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc

pprStmtInCtxt :: (OutputableBndrId idL, OutputableBndrId idR, Outputable body, Anno (StmtLR (GhcPass idL) (GhcPass idR) body) ~ SrcSpanAnnA) => HsStmtContext (GhcPass idL) -> StmtLR (GhcPass idL) (GhcPass idR) body -> SDoc

pprTransStmt :: Outputable body => Maybe body -> body -> TransForm -> SDoc

pprTransformStmt :: OutputableBndrId p => [IdP (GhcPass p)] -> LHsExpr (GhcPass p) -> Maybe (LHsExpr (GhcPass p)) -> SDoc

ppr_apps :: OutputableBndrId p => HsExpr (GhcPass p) -> [Either (LHsExpr (GhcPass p)) (LHsWcType (NoGhcTc (GhcPass p)))] -> SDoc

ppr_cmd :: forall p. OutputableBndrId p => HsCmd (GhcPass p) -> SDoc

ppr_do_stmts :: (OutputableBndrId idL, OutputableBndrId idR, Anno (StmtLR (GhcPass idL) (GhcPass idR) body) ~ SrcSpanAnnA, Outputable body) => [LStmtLR (GhcPass idL) (GhcPass idR) body] -> SDoc

ppr_expr :: forall p. OutputableBndrId p => HsExpr (GhcPass p) -> SDoc

ppr_infix_expr :: forall p. OutputableBndrId p => HsExpr (GhcPass p) -> Maybe SDoc

ppr_lcmd :: OutputableBndrId p => LHsCmd (GhcPass p) -> SDoc

ppr_lexpr :: OutputableBndrId p => LHsExpr (GhcPass p) -> SDoc

ppr_module_name_prefix :: Maybe ModuleName -> SDoc

ppr_quasi :: OutputableBndr p => p -> p -> FastString -> SDoc

ppr_splice :: OutputableBndrId p => SDoc -> IdP (GhcPass p) -> LHsExpr (GhcPass p) -> SDoc -> SDoc

ppr_splice_decl :: OutputableBndrId p => HsSplice (GhcPass p) -> SDoc

stripParensHsExpr :: HsExpr (GhcPass p) -> HsExpr (GhcPass p)

stripParensLHsExpr :: LHsExpr (GhcPass p) -> LHsExpr (GhcPass p)

thBrackets :: SDoc -> SDoc -> SDoc

thTyBrackets :: SDoc -> SDoc

tupArgPresent :: HsTupArg (GhcPass p) -> Bool

data AnnExplicitSum

Constructors

AnnExplicitSum

Fields

aesOpen :: EpaLocation

aesBarsBefore :: [EpaLocation]

aesBarsAfter :: [EpaLocation]

aesClose :: EpaLocation

data AnnFieldLabel

Constructors

AnnFieldLabel

Fields

afDot :: Maybe EpaLocation

data AnnProjection

Constructors

AnnProjection

Fields

apOpen :: EpaLocation

'('

apClose :: EpaLocation

')'

data AnnsIf

Constructors

AnnsIf

Fields

aiIf :: EpaLocation

aiThen :: EpaLocation

aiElse :: EpaLocation

aiThenSemi :: Maybe EpaLocation

aiElseSemi :: Maybe EpaLocation

data AnnsLet

Constructors

AnnsLet

Fields

alLet :: EpaLocation

alIn :: EpaLocation

data CmdTopTc

Constructors

CmdTopTc Type Type (CmdSyntaxTable GhcTc)

data DelayedSplice

Constructors

DelayedSplice TcLclEnv (LHsExpr GhcRn) TcType (LHsExpr GhcTc)

data EpAnnHsCase

Constructors

EpAnnHsCase

Fields

hsCaseAnnCase :: EpaLocation

hsCaseAnnOf :: EpaLocation

hsCaseAnnsRest :: [AddEpAnn]

data EpAnnUnboundVar

Constructors

EpAnnUnboundVar

Fields

hsUnboundBackquotes :: (EpaLocation, EpaLocation)

hsUnboundHole :: EpaLocation

data GrhsAnn

Constructors

GrhsAnn

Fields

ga_vbar :: Maybe EpaLocation

ga_sep :: AddEpAnn

Match separator location

data HsExpansion a b

Constructors

HsExpanded a b

newtype HsSplicedT

Constructors

HsSplicedT DelayedSplice

data HsWrap hs_syn

HsWrap appears only in typechecker output Invariant: The contained Expr is *NOT* itself an HsWrap. See Note [Detecting forced eta expansion] in GHC.HsToCore.Expr. This invariant is maintained by mkHsWrap. hs_syn is something like HsExpr or HsCmd

Constructors

HsWrap HsWrapper (hs_syn GhcTc)

data PendingRnSplice

Pending Renamer Splice

Constructors

PendingRnSplice UntypedSpliceFlavour SplicePointName (LHsExpr GhcRn)

data PendingTcSplice

Pending Type-checker Splice

Constructors

PendingTcSplice SplicePointName (LHsExpr GhcTc)

type PostTcExpr = HsExpr GhcTc

Post-Type checking Expression

PostTcExpr is an evidence expression attached to the syntax tree by the type checker (c.f. postTcType).

type PostTcTable = [(Name, PostTcExpr)]

Post-Type checking Table

We use a PostTcTable where there are a bunch of pieces of evidence, more than is convenient to keep individually.

data RecStmtTc

Constructors

RecStmtTc

Fields

recS_bind_ty :: Type

recS_later_rets :: [PostTcExpr]

recS_rec_rets :: [PostTcExpr]

recS_ret_ty :: Type

data RecordUpdTc

Extra data fields for a RecordUpd, added by the type checker

Constructors

RecordUpdTc

Fields

rupd_cons :: [ConLike]

rupd_in_tys :: [Type]

rupd_out_tys :: [Type]

rupd_wrap :: HsWrapper

type family SyntaxExprGhc (p :: Pass) = (r :: Type) | r -> p where ...

Equations

SyntaxExprGhc 'Parsed = NoExtField

SyntaxExprGhc 'Renamed = SyntaxExprRn

SyntaxExprGhc 'Typechecked = SyntaxExprTc

data SyntaxExprRn

The function to use in rebindable syntax. See Note [NoSyntaxExpr].

Constructors

SyntaxExprRn (HsExpr GhcRn)

NoSyntaxExprRn

data SyntaxExprTc

An expression with wrappers, used for rebindable syntax

This should desugar to

syn_res_wrap $ syn_expr (syn_arg_wraps[0] arg0)
                        (syn_arg_wraps[1] arg1) ...

where the actual arguments come from elsewhere in the AST.

Constructors

SyntaxExprTc

Fields

syn_expr :: HsExpr GhcTc

syn_arg_wraps :: [HsWrapper]

syn_res_wrap :: HsWrapper

NoSyntaxExprTc

data XBindStmtRn

Constructors

XBindStmtRn

Fields

xbsrn_bindOp :: SyntaxExpr GhcRn

xbsrn_failOp :: FailOperator GhcRn

data XBindStmtTc

Constructors

XBindStmtTc

Fields

xbstc_bindOp :: SyntaxExpr GhcTc

xbstc_boundResultType :: Type

xbstc_boundResultMult :: Mult

xbstc_failOp :: FailOperator GhcTc

data XXExprGhcTc

Constructors

WrapExpr !(HsWrap HsExpr)

ExpansionExpr !(HsExpansion (HsExpr GhcRn) (HsExpr GhcTc))

Orphan instances

Outputable HsArrowMatchContext

Instance details

Methods

ppr :: HsArrowMatchContext -> SDoc

OutputableBndrId idL => Outputable (ApplicativeArg (GhcPass idL))

Instance details

Methods

ppr :: ApplicativeArg (GhcPass idL) -> SDoc

OutputableBndrId p => Outputable (ArithSeqInfo (GhcPass p))

Instance details

Methods

ppr :: ArithSeqInfo (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsBracket (GhcPass p))

Instance details

Methods

ppr :: HsBracket (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsCmd (GhcPass p))

Instance details

Methods

ppr :: HsCmd (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsCmdTop (GhcPass p))

Instance details

Methods

ppr :: HsCmdTop (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsExpr (GhcPass p))

Instance details

Methods

ppr :: HsExpr (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsMatchContext (GhcPass p))

Instance details

Methods

ppr :: HsMatchContext (GhcPass p) -> SDoc

Outputable (HsPragE (GhcPass p))

Instance details

Methods

ppr :: HsPragE (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsSplice (GhcPass p))

Instance details

Methods

ppr :: HsSplice (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsSplicedThing (GhcPass p))

Instance details

Methods

ppr :: HsSplicedThing (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsStmtContext (GhcPass p))

Instance details

Methods

ppr :: HsStmtContext (GhcPass p) -> SDoc

Anno a ~ SrcSpanAnn' (EpAnn an) => WrapXRec (GhcPass p) a

Instance details

Methods

wrapXRec :: a -> XRec (GhcPass p) a

(OutputableBndrId pr, Outputable body) => Outputable (Match (GhcPass pr) body)

Instance details

Methods

ppr :: Match (GhcPass pr) body -> SDoc

(Outputable (StmtLR (GhcPass idL) (GhcPass idL) (LHsExpr (GhcPass idL))), Outputable (XXParStmtBlock (GhcPass idL) (GhcPass idR))) => Outputable (ParStmtBlock (GhcPass idL) (GhcPass idR))

Instance details

Methods

ppr :: ParStmtBlock (GhcPass idL) (GhcPass idR) -> SDoc

(OutputableBndrId pl, OutputableBndrId pr, Anno (StmtLR (GhcPass pl) (GhcPass pr) body) ~ SrcSpanAnnA, Outputable body) => Outputable (StmtLR (GhcPass pl) (GhcPass pr) body)

Instance details

Methods

ppr :: StmtLR (GhcPass pl) (GhcPass pr) body -> SDoc