GHC.Hs.ImpExp

Documentation

type LImportDecl pass

Arguments

= XRec pass (ImportDecl pass)
When in a list this may have

AnnKeywordId : AnnSemi

Located Import Declaration

data ImportDeclQualifiedStyle

If/how an import is qualified.

Constructors

QualifiedPre
qualified appears in prepositive position.

QualifiedPost
qualified appears in postpositive position.

NotQualified
Not qualified.

importDeclQualifiedStyle :: Maybe EpaLocation -> Maybe EpaLocation -> (Maybe EpaLocation, ImportDeclQualifiedStyle)

Given two possible located qualified tokens, compute a style (in a conforming Haskell program only one of the two can be not Nothing). This is called from GHC.Parser.

isImportDeclQualified :: ImportDeclQualifiedStyle -> Bool

Convenience function to answer the question if an import decl. is qualified.

data ImportDecl pass

Import Declaration

A single Haskell import declaration.

Constructors

ImportDecl

Fields

ideclExt :: XCImportDecl pass

ideclSourceSrc :: SourceText

ideclName :: XRec pass ModuleName

Module name.

ideclPkgQual :: Maybe StringLiteral

Package qualifier.

ideclSource :: IsBootInterface

IsBoot = {-# SOURCE #-} import

ideclSafe :: Bool

True => safe import

ideclQualified :: ImportDeclQualifiedStyle

If/how the import is qualified.

ideclImplicit :: Bool

True => implicit import (of Prelude)

ideclAs :: Maybe (XRec pass ModuleName)

as Module

ideclHiding :: Maybe (Bool, XRec pass [LIE pass])

(True => hiding, names)

XImportDecl !(XXImportDecl pass)
AnnKeywordIds

AnnImport

AnnOpen, AnnClose for ideclSource

AnnSafe,AnnQualified, AnnPackageName,AnnAs, AnnVal

AnnHiding,AnnOpen, AnnClose attached to location in ideclHiding

data EpAnnImportDecl

Constructors

EpAnnImportDecl

Fields

importDeclAnnImport :: EpaLocation

importDeclAnnPragma :: Maybe (EpaLocation, EpaLocation)

importDeclAnnSafe :: Maybe EpaLocation

importDeclAnnQualified :: Maybe EpaLocation

importDeclAnnPackage :: Maybe EpaLocation

importDeclAnnAs :: Maybe EpaLocation

simpleImportDecl :: ModuleName -> ImportDecl GhcPs

data IEWrappedName name

A name in an import or export specification which may have adornments. Used primarily for accurate pretty printing of ParsedSource, and API Annotation placement. The Annotation is the location of the adornment in the original source.

Constructors

IEName (LocatedN name)
no extra

IEPattern EpaLocation (LocatedN name)
pattern X

IEType EpaLocation (LocatedN name)
type (:+:)

type LIEWrappedName name = LocatedA (IEWrappedName name)

Located name with possible adornment - AnnKeywordIds : AnnType, AnnPattern

type LIE pass

Arguments

= XRec pass (IE pass)
When in a list this may have

AnnKeywordId : AnnComma

Located Import or Export

data IE pass

Imported or exported entity.

Constructors

IEVar (XIEVar pass) (LIEWrappedName (IdP pass))
Imported or Exported Variable

IEThingAbs (XIEThingAbs pass) (LIEWrappedName (IdP pass))
Imported or exported Thing with Absent list

The thing is a Class/Type (can't tell) - AnnKeywordIds : AnnPattern, AnnType,AnnVal

IEThingAll (XIEThingAll pass) (LIEWrappedName (IdP pass))
Imported or exported Thing with All imported or exported

The thing is a ClassType and the All refers to methodsconstructors

AnnKeywordIds : AnnOpen, AnnDotdot,AnnClose, AnnType

IEThingWith (XIEThingWith pass) (LIEWrappedName (IdP pass)) IEWildcard [LIEWrappedName (IdP pass)]
Imported or exported Thing With given imported or exported

The thing is a Class/Type and the imported or exported things are methods/constructors and record fields; see Note [IEThingWith] - AnnKeywordIds : AnnOpen, AnnClose, AnnComma, AnnType

IEModuleContents (XIEModuleContents pass) (XRec pass ModuleName)
Imported or exported module contents

(Export Only)

AnnKeywordIds : AnnModule

IEGroup (XIEGroup pass) Int HsDocString
Doc section heading

IEDoc (XIEDoc pass) HsDocString
Some documentation

IEDocNamed (XIEDocNamed pass) String
Reference to named doc

XIE !(XXIE pass)

data IEWildcard

Imported or Exported Wildcard

Constructors

NoIEWildcard

IEWildcard Int

ieName :: IE (GhcPass p) -> IdP (GhcPass p)

ieNames :: IE (GhcPass p) -> [IdP (GhcPass p)]

ieWrappedLName :: IEWrappedName name -> LocatedN name

ieWrappedName :: IEWrappedName name -> name

lieWrappedName :: LIEWrappedName name -> name

ieLWrappedName :: LIEWrappedName name -> LocatedN name

replaceWrappedName :: IEWrappedName name1 -> name2 -> IEWrappedName name2

replaceLWrappedName :: LIEWrappedName name1 -> name2 -> LIEWrappedName name2

pprImpExp :: (HasOccName name, OutputableBndr name) => name -> SDoc