GHC.Hs.Type

Documentation

type Mult = Type

Mult is a type alias for Type.

Mult must contain Type because multiplicity variables are mere type variables (of kind Multiplicity) in Haskell. So the simplest implementation is to make Mult be Type.

Multiplicities can be formed with: - One: GHC.Types.One (= oneDataCon) - Many: GHC.Types.Many (= manyDataCon) - Multiplication: GHC.Types.MultMul (= multMulTyCon)

So that Mult feels a bit more structured, we provide pattern synonyms and smart constructors for these.

data HsScaled pass a

This is used in the syntax. In constructor declaration. It must keep the arrow representation.

Constructors

HsScaled (HsArrow pass) a

hsMult :: HsScaled pass a -> HsArrow pass

hsScaledThing :: HsScaled pass a -> a

type HsArrow pass = HsArrowOf (LHsType pass) pass

data HsArrowOf mult pass

Denotes the type of arrows in the surface language

Constructors

HsUnrestrictedArrow !(XUnrestrictedArrow mult pass)
a -> b or a → b

HsLinearArrow !(XLinearArrow mult pass)
a %1 -> b or a %1 → b, or a ⊸ b

HsExplicitMult !(XExplicitMult mult pass) !mult
a %m -> b or a %m → b (very much including `a %Many -> b`! This is how the programmer wrote it). It is stored as an HsType so as to preserve the syntax as written in the program.

XArrow !(XXArrow mult pass)

arrowToHsType :: HsArrow GhcRn -> LHsType GhcRn

expandHsArrow :: (LocatedN Name -> t GhcRn) -> HsArrowOf (LocatedA (t GhcRn)) GhcRn -> LocatedA (t GhcRn)

Convert an arrow into its corresponding multiplicity. In essence this erases the information of whether the programmer wrote an explicit multiplicity or a shorthand.

data EpLinearArrow

Constructors

EpPct1 !(EpToken "%1") !TokRarrow

EpLolly !(EpToken "\8888")

hsLinear :: forall (p :: Pass) a. IsPass p => a -> HsScaled (GhcPass p) a

hsUnrestricted :: forall (p :: Pass) a. IsPass p => a -> HsScaled (GhcPass p) a

isUnrestricted :: HsArrow GhcRn -> Bool

pprHsArrow :: forall mult (pass :: Pass). (Outputable mult, OutputableBndrId pass) => HsArrowOf mult (GhcPass pass) -> SDoc

data HsType pass

Haskell Type

Constructors

HsForAllTy

Fields

hst_xforall :: XForAllTy pass

hst_tele :: HsForAllTelescope pass

hst_body :: LHsType pass

HsQualTy

Fields

hst_xqual :: XQualTy pass

hst_ctxt :: LHsContext pass

hst_body :: LHsType pass

HsTyVar (XTyVar pass) PromotionFlag (LIdP pass)

HsAppTy (XAppTy pass) (LHsType pass) (LHsType pass)

HsAppKindTy (XAppKindTy pass) (LHsType pass) (LHsKind pass)

HsFunTy (XFunTy pass) (HsArrow pass) (LHsType pass) (LHsType pass)

HsListTy (XListTy pass) (LHsType pass)

HsTupleTy (XTupleTy pass) HsTupleSort [LHsType pass]

HsSumTy (XSumTy pass) [LHsType pass]

HsOpTy (XOpTy pass) PromotionFlag (LHsType pass) (LIdP pass) (LHsType pass)

HsParTy (XParTy pass) (LHsType pass)

HsIParamTy (XIParamTy pass) (XRec pass HsIPName) (LHsType pass)
(?x :: ty)

HsStarTy (XStarTy pass) Bool

HsKindSig (XKindSig pass) (LHsType pass) (LHsKind pass)
(ty :: kind)

HsSpliceTy (XSpliceTy pass) (HsUntypedSplice pass)

HsDocTy (XDocTy pass) (LHsType pass) (LHsDoc pass)

HsBangTy (XBangTy pass) HsBang (LHsType pass)

HsRecTy (XRecTy pass) [LConDeclField pass]

HsExplicitListTy (XExplicitListTy pass) PromotionFlag [LHsType pass]

HsExplicitTupleTy (XExplicitTupleTy pass) PromotionFlag [LHsType pass]

HsTyLit (XTyLit pass) (HsTyLit pass)

HsWildCardTy (XWildCardTy pass)

XHsType !(XXType pass)

type HsCoreTy = Type

type LHsType pass = XRec pass (HsType pass)

Located Haskell Type

type HsKind pass = HsType pass

Haskell Kind

type LHsKind pass = XRec pass (HsKind pass)

Located Haskell Kind

data HsForAllTelescope pass

The type variable binders in an HsForAllTy. See also Note [Variable Specificity and Forall Visibility] in GHC.Tc.Gen.HsType.

Constructors

HsForAllVis
A visible forall (e.g., forall a -> {...}). These do not have any notion of specificity, so we use () as a placeholder value.

Fields

hsf_xvis :: XHsForAllVis pass

hsf_vis_bndrs :: [LHsTyVarBndr () pass]

HsForAllInvis
An invisible forall (e.g., forall a {b} c. {...}), where each binder has a Specificity.

Fields

hsf_xinvis :: XHsForAllInvis pass

hsf_invis_bndrs :: [LHsTyVarBndr Specificity pass]

XHsForAllTelescope !(XXHsForAllTelescope pass)

type EpAnnForallVis = EpAnn (TokForall, TokRarrow)

type EpAnnForallInvis = EpAnn (TokForall, EpToken ".")

data HsTyVarBndr flag pass

Haskell Type Variable Binder See Note [Type variable binders]

Constructors

HsTvb

Fields

tvb_ext :: XTyVarBndr pass

tvb_flag :: flag

tvb_var :: HsBndrVar pass

tvb_kind :: HsBndrKind pass

XTyVarBndr !(XXTyVarBndr pass)

type LHsTyVarBndr flag pass = XRec pass (HsTyVarBndr flag pass)

Located Haskell Type Variable Binder

data AnnTyVarBndr

Constructors

AnnTyVarBndr

Fields

atv_opens :: [EpaLocation]

atv_closes :: [EpaLocation]

atv_tv :: EpToken "'"

atv_dcolon :: TokDcolon

data HsBndrKind pass

Constructors

HsBndrKind !(XBndrKind pass) (LHsKind pass)

HsBndrNoKind !(XBndrNoKind pass)

XBndrKind !(XXBndrKind pass)

data HsBndrVar pass

Constructors

HsBndrVar !(XBndrVar pass) !(LIdP pass)

HsBndrWildCard !(XBndrWildCard pass)

XBndrVar !(XXBndrVar pass)

data HsBndrVis pass

Constructors

HsBndrRequired !(XBndrRequired pass)

HsBndrInvisible !(XBndrInvisible pass)

XBndrVis !(XXBndrVis pass)

isHsBndrInvisible :: HsBndrVis pass -> Bool

data LHsQTyVars pass

Located Haskell Quantified Type Variables

Constructors

HsQTvs

Fields

hsq_ext :: XHsQTvs pass

hsq_explicit :: [LHsTyVarBndr (HsBndrVis pass) pass]

XLHsQTyVars !(XXLHsQTyVars pass)

data HsOuterTyVarBndrs flag pass

The outermost type variables in a type that obeys the forall-or-nothing rule. See Note [forall-or-nothing rule].

Constructors

HsOuterImplicit
Implicit forall, e.g., f :: a -> b -> b

Fields

hso_ximplicit :: XHsOuterImplicit pass

HsOuterExplicit
Explicit forall, e.g., f :: forall a b. a -> b -> b

Fields

hso_xexplicit :: XHsOuterExplicit pass flag

hso_bndrs :: [LHsTyVarBndr flag (NoGhcTc pass)]

XHsOuterTyVarBndrs !(XXHsOuterTyVarBndrs pass)

type HsOuterFamEqnTyVarBndrs = HsOuterTyVarBndrs ()

Used for type-family instance equations, e.g.,

type instance forall a. F [a] = Tree a

The notion of specificity is irrelevant in type family equations, so we use () for the HsOuterTyVarBndrs flag.

type HsOuterSigTyVarBndrs = HsOuterTyVarBndrs Specificity

Used for signatures, e.g.,

f :: forall a {b}. blah

We use Specificity for the HsOuterTyVarBndrs flag to allow distinguishing between specified and inferred type variables.

data HsWildCardBndrs pass thing

Haskell Wildcard Binders

Constructors

HsWC

Fields

hswc_ext :: XHsWC pass thing

hswc_body :: thing

XHsWildCardBndrs !(XXHsWildCardBndrs pass thing)

data HsPatSigType pass

Types that can appear in pattern signatures, as well as the signatures for term-level binders in RULES. See Note [Pattern signature binders and scoping].

This is very similar to HsSigWcType, but with slightly different semantics: see Note [HsType binders]. See also Note [The wildcard story for types].

Constructors

HsPS

Fields

hsps_ext :: XHsPS pass

After renamer: HsPSRn

hsps_body :: LHsType pass

Main payload (the type itself)

XHsPatSigType !(XXHsPatSigType pass)

data HsPSRn

The extension field for HsPatSigType, which is only used in the renamer onwards. See Note [Pattern signature binders and scoping].

Constructors

HsPSRn

Fields

hsps_nwcs :: [Name]

Wildcard names

hsps_imp_tvs :: [Name]

Implicitly bound variable names

data HsTyPat pass

Constructors

HsTP

Fields

hstp_ext :: XHsTP pass

After renamer: HsTyPatRn

hstp_body :: LHsType pass

Main payload (the type itself)

XHsTyPat !(XXHsTyPat pass)

data HsTyPatRn

Constructors

HsTPRn

Fields

hstp_nwcs :: [Name]

Wildcard names

hstp_imp_tvs :: [Name]

Implicitly bound variable names

hstp_exp_tvs :: [Name]

Explicitly bound variable names

data HsTyPatRnBuilder

A variant of HsTyPatRn that uses Bags for efficient concatenation. See Note [Implicit and explicit type variable binders] in GHC.Rename.Pat

Constructors

HsTPRnB

Fields

hstpb_nwcs :: Bag Name

hstpb_imp_tvs :: Bag Name

hstpb_exp_tvs :: Bag Name

tpBuilderExplicitTV :: Name -> HsTyPatRnBuilder

tpBuilderPatSig :: HsPSRn -> HsTyPatRnBuilder

buildHsTyPatRn :: HsTyPatRnBuilder -> HsTyPatRn

builderFromHsTyPatRn :: HsTyPatRn -> HsTyPatRnBuilder

data HsSigType pass

A type signature that obeys the forall-or-nothing rule. In other words, an LHsType that uses an HsOuterSigTyVarBndrs to represent its outermost type variable quantification. See Note [Representing type signatures].

Constructors

HsSig

Fields

sig_ext :: XHsSig pass

sig_bndrs :: HsOuterSigTyVarBndrs pass

sig_body :: LHsType pass

XHsSigType !(XXHsSigType pass)

type LHsSigType pass = XRec pass (HsSigType pass)

Located Haskell Signature Type

type LHsSigWcType pass = HsWildCardBndrs pass (LHsSigType pass)

Located Haskell Signature Wildcard Type

type LHsWcType pass = HsWildCardBndrs pass (LHsType pass)

Located Haskell Wildcard Type

data HsTupleSort

Haskell Tuple Sort

Constructors

HsUnboxedTuple

HsBoxedOrConstraintTuple

type HsContext pass = [LHsType pass]

Haskell Context

type LHsContext pass = XRec pass (HsContext pass)

Located Haskell Context

fromMaybeContext :: forall (p :: Pass). Maybe (LHsContext (GhcPass p)) -> HsContext (GhcPass p)

data HsTyLit pass

Haskell Type Literal

Constructors

HsNumTy (XNumTy pass) Integer

HsStrTy (XStrTy pass) FastString

HsCharTy (XCharTy pass) Char

XTyLit !(XXTyLit pass)

newtype HsIPName

These names are used early on to store the names of implicit parameters. They completely disappear after type-checking.

Constructors

HsIPName FastString

hsIPNameFS :: HsIPName -> FastString

data HsArg p tm ty

Arguments in an expression/type after splitting

Constructors

HsValArg !(XValArg p) tm

HsTypeArg !(XTypeArg p) ty

HsArgPar !(XArgPar p)

XArg !(XXArg p)

numVisibleArgs :: [HsArg p tm ty] -> Arity

pprHsArgsApp :: forall id tm ty (p :: Pass). (OutputableBndr id, Outputable tm, Outputable ty) => id -> LexicalFixity -> [HsArg (GhcPass p) tm ty] -> SDoc

pprHsArgsApp id fixity args pretty-prints an application of id to args, using the fixity to tell whether id should be printed prefix or infix. Examples:

pprHsArgsApp T Prefix [HsTypeArg Bool, HsValArg Int]                        = T @Bool Int
pprHsArgsApp T Prefix [HsTypeArg Bool, HsArgPar, HsValArg Int]              = (T @Bool) Int
pprHsArgsApp (++) Infix [HsValArg Char, HsValArg Double]                    = Char ++ Double
pprHsArgsApp (++) Infix [HsValArg Char, HsValArg Double, HsVarArg Ordering] = (Char ++ Double) Ordering

type LHsTypeArg p = HsArg p (LHsType p) (LHsKind p)

lhsTypeArgSrcSpan :: LHsTypeArg GhcPs -> SrcSpan

Compute the SrcSpan associated with an LHsTypeArg.

class OutputableBndrFlag flag (p :: Pass)

Minimal complete definition

pprTyVarBndr

type LBangType pass = XRec pass (BangType pass)

Located Bang Type

type BangType pass = HsType pass

Bang Type

In the parser, strictness and packedness annotations bind more tightly than docstrings. This means that when consuming a BangType (and looking for HsBangTy) we must be ready to peer behind a potential layer of HsDocTy. See #15206 for motivation and getBangType for an example.

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as written by the user, including the source code for exact-printing.

In the AST, the SourceText is deconstructed and hidden inside XBangTy extension point.

Constructors

HsSrcBang SourceText HsBang

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field, or one with an unlifted type

HsStrict Bool
Strict but not unpacked field True = we could have unpacked, but opted not to because of -O0. See Note [Detecting useless UNPACK pragmas]

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie ~

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{-# UNPACK #-} specified

SrcNoUnpack
{-# NOUNPACK #-} specified

NoSrcUnpack
no unpack pragma

getBangType :: forall (p :: Pass). LHsType (GhcPass p) -> LHsType (GhcPass p)

getBangStrictness :: forall (p :: Pass). LHsType (GhcPass p) -> HsSrcBang

data ConDeclField pass

Constructor Declaration Field

Constructors

ConDeclField

Fields

cd_fld_ext :: XConDeclField pass

cd_fld_names :: [LFieldOcc pass]

See Note [ConDeclField pass]

cd_fld_type :: LBangType pass

cd_fld_doc :: Maybe (LHsDoc pass)

XConDeclField !(XXConDeclField pass)

type LConDeclField pass = XRec pass (ConDeclField pass)

Located Constructor Declaration Field

pprConDeclFields :: forall (p :: Pass). OutputableBndrId p => [LConDeclField (GhcPass p)] -> SDoc

data HsConDetails tyarg arg rec

Describes the arguments to a data constructor. This is a common representation for several constructor-related concepts, including:

The arguments in a Haskell98-style constructor declaration (see HsConDeclH98Details in GHC.Hs.Decls).

The arguments in constructor patterns in case/function definitions (see HsConPatDetails in GHC.Hs.Pat).

The left-hand side arguments in a pattern synonym binding (see HsPatSynDetails in GHC.Hs.Binds).

One notable exception is the arguments in a GADT constructor, which uses a separate data type entirely (see HsConDeclGADTDetails in GHC.Hs.Decls). This is because GADT constructors cannot be declared with infix syntax, unlike the concepts above (#18844).

Constructors

PrefixCon [tyarg] [arg]

RecCon rec

InfixCon arg arg

noTypeArgs :: [Void]

An empty list that can be used to indicate that there are no type arguments allowed in cases where HsConDetails is applied to Void.

data FieldOcc pass

Field Occurrence

Represents an *occurrence* of a field. This may or may not be a binding occurrence (e.g. this type is used in ConDeclField and RecordPatSynField which bind their fields, but also in HsRecField for record construction and patterns, which do not).

We store both the RdrName the user originally wrote, and after the renamer we use the extension field to store the selector function.

There is a wrinkle in that update field occurances are sometimes ambiguous during the rename stage. See note [Ambiguous FieldOcc in record updates] to see how we currently handle this.

Constructors

FieldOcc

Fields

foExt :: XCFieldOcc pass

foLabel :: LIdP pass

XFieldOcc !(XXFieldOcc pass)

type LFieldOcc pass = XRec pass (FieldOcc pass)

Located Field Occurrence

mkFieldOcc :: LocatedN RdrName -> FieldOcc GhcPs

fieldOccRdrName :: forall (p :: Pass). IsPass p => FieldOcc (GhcPass p) -> RdrName

fieldOccLRdrName :: forall (p :: Pass). IsPass p => FieldOcc (GhcPass p) -> LocatedN RdrName

data OpName

Name of an operator in an operator application or section

Constructors

NormalOp Name
A normal identifier

NegateOp
Prefix negation

UnboundOp RdrName
An unbound identifier

RecFldOp (FieldOcc GhcRn)
A record field occurrence

mkAnonWildCardTy :: EpToken "_" -> HsType GhcPs

pprAnonWildCard :: SDoc

hsOuterTyVarNames :: HsOuterTyVarBndrs flag GhcRn -> [Name]

hsOuterExplicitBndrs :: forall flag (p :: Pass). HsOuterTyVarBndrs flag (GhcPass p) -> [LHsTyVarBndr flag (NoGhcTc (GhcPass p))]

mapHsOuterImplicit :: (XHsOuterImplicit pass -> XHsOuterImplicit pass) -> HsOuterTyVarBndrs flag pass -> HsOuterTyVarBndrs flag pass

mkHsOuterImplicit :: HsOuterTyVarBndrs flag GhcPs

mkHsOuterExplicit :: EpAnnForallInvis -> [LHsTyVarBndr flag GhcPs] -> HsOuterTyVarBndrs flag GhcPs

mkHsImplicitSigType :: LHsType GhcPs -> HsSigType GhcPs

mkHsExplicitSigType :: EpAnnForallInvis -> [LHsTyVarBndr Specificity GhcPs] -> LHsType GhcPs -> HsSigType GhcPs

mkHsWildCardBndrs :: thing -> HsWildCardBndrs GhcPs thing

mkHsPatSigType :: EpAnnCO -> LHsType GhcPs -> HsPatSigType GhcPs

mkHsTyPat :: LHsType GhcPs -> HsTyPat GhcPs

mkEmptyWildCardBndrs :: thing -> HsWildCardBndrs GhcRn thing

mkHsForAllVisTele :: forall (p :: Pass). EpAnnForallVis -> [LHsTyVarBndr () (GhcPass p)] -> HsForAllTelescope (GhcPass p)

mkHsForAllInvisTele :: forall (p :: Pass). EpAnnForallInvis -> [LHsTyVarBndr Specificity (GhcPass p)] -> HsForAllTelescope (GhcPass p)

mkHsQTvs :: [LHsTyVarBndr (HsBndrVis GhcPs) GhcPs] -> LHsQTyVars GhcPs

hsQTvExplicit :: LHsQTyVars pass -> [LHsTyVarBndr (HsBndrVis pass) pass]

emptyLHsQTvs :: LHsQTyVars GhcRn

isHsKindedTyVar :: HsTyVarBndr flag pass -> Bool

Does this HsTyVarBndr come with an explicit kind annotation?

hsBndrVar :: forall flag (pass :: Pass). HsTyVarBndr flag (GhcPass pass) -> HsBndrVar (GhcPass pass)

Get the variable of the type variable binder

hsBndrKind :: forall flag (pass :: Pass). HsTyVarBndr flag (GhcPass pass) -> HsBndrKind (GhcPass pass)

Get the kind of the type variable binder

hsTvbAllKinded :: forall (p :: Pass). LHsQTyVars (GhcPass p) -> Bool

Do all type variables in this LHsQTyVars come with kind annotations?

hsScopedTvs :: LHsSigType GhcRn -> [Name]

hsScopedKvs :: LHsKind GhcRn -> [Name]

hsWcScopedTvs :: LHsSigWcType GhcRn -> [Name]

dropWildCards :: LHsSigWcType pass -> LHsSigType pass

hsTyVarLName :: forall flag (p :: Pass). HsTyVarBndr flag (GhcPass p) -> Maybe (LIdP (GhcPass p))

hsTyVarName :: forall flag (p :: Pass). HsTyVarBndr flag (GhcPass p) -> Maybe (IdP (GhcPass p))

hsAllLTyVarNames :: LHsQTyVars GhcRn -> [Name]

hsLTyVarLocNames :: forall (p :: Pass). Anno (IdGhcP p) ~ SrcSpanAnnN => LHsQTyVars (GhcPass p) -> [LocatedN (IdP (GhcPass p))]

hsLTyVarName :: forall flag (p :: Pass). LHsTyVarBndr flag (GhcPass p) -> Maybe (IdP (GhcPass p))

hsLTyVarNames :: forall flag (p :: Pass). [LHsTyVarBndr flag (GhcPass p)] -> [IdP (GhcPass p)]

hsForAllTelescopeNames :: forall (p :: Pass). HsForAllTelescope (GhcPass p) -> [IdP (GhcPass p)]

hsLTyVarLocName :: forall (p :: Pass) flag. Anno (IdGhcP p) ~ SrcSpanAnnN => LHsTyVarBndr flag (GhcPass p) -> Maybe (LocatedN (IdP (GhcPass p)))

hsExplicitLTyVarNames :: forall (p :: Pass). LHsQTyVars (GhcPass p) -> [IdP (GhcPass p)]

splitLHsInstDeclTy :: LHsSigType GhcRn -> ([Name], Maybe (LHsContext GhcRn), LHsType GhcRn)

Decompose a type class instance type (of the form forall tvs. context => instance_head) into its constituent parts. Note that the [Name]s returned correspond to either:

The implicitly bound type variables (if the type lacks an outermost forall), or

The explicitly bound type variables (if the type has an outermost forall).

This function is careful not to look through parentheses. See Note [No nested foralls or contexts in instance types] for why this is important.

getLHsInstDeclHead :: forall (p :: Pass). LHsSigType (GhcPass p) -> LHsType (GhcPass p)

Decompose a type class instance type (of the form forall tvs. context => instance_head) into the instance_head.

getLHsInstDeclClass_maybe :: forall (p :: Pass). Anno (IdGhcP p) ~ SrcSpanAnnN => LHsSigType (GhcPass p) -> Maybe (LocatedN (IdP (GhcPass p)))

Decompose a type class instance type (of the form forall tvs. context => instance_head) into the instance_head and retrieve the underlying class type constructor (if it exists).

splitLHsPatSynTy :: forall (p :: Pass). LHsSigType (GhcPass p) -> ([LHsTyVarBndr Specificity (GhcPass (NoGhcTcPass p))], Maybe (LHsContext (GhcPass p)), [LHsTyVarBndr Specificity (GhcPass p)], Maybe (LHsContext (GhcPass p)), LHsType (GhcPass p))

Decompose a pattern synonym type signature into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsForAllTyInvis :: forall (pass :: Pass). LHsType (GhcPass pass) -> ([LHsTyVarBndr Specificity (GhcPass pass)], LHsType (GhcPass pass))

Decompose a type of the form forall tvs. body into its constituent parts. Only splits type variable binders that were quantified invisibly (e.g., forall a., with a dot).

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them. Unlike splitLHsSigmaTyInvis, this function does not look through parentheses, hence the suffix _KP (short for "Keep Parentheses").

splitLHsForAllTyInvis_KP :: forall (pass :: Pass). LHsType (GhcPass pass) -> (Maybe [LHsTyVarBndr Specificity (GhcPass pass)], LHsType (GhcPass pass))

Decompose a type of the form forall tvs. body into its constituent parts. Only splits type variable binders that were quantified invisibly (e.g., forall a., with a dot).

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Unlike splitLHsForAllTyInvis, this function does not look through parentheses, hence the suffix _KP (short for "Keep Parentheses").

splitLHsQualTy :: forall (pass :: Pass). LHsType (GhcPass pass) -> (Maybe (LHsContext (GhcPass pass)), LHsType (GhcPass pass))

Decompose a type of the form context => body into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (context => ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsSigmaTyInvis :: forall (p :: Pass). LHsType (GhcPass p) -> ([LHsTyVarBndr Specificity (GhcPass p)], Maybe (LHsContext (GhcPass p)), LHsType (GhcPass p))

Decompose a sigma type (of the form forall tvs. context => body) into its constituent parts. Only splits type variable binders that were quantified invisibly (e.g., forall a., with a dot).

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsGadtTy :: LHsSigType GhcPs -> (HsOuterSigTyVarBndrs GhcPs, Maybe (LHsContext GhcPs), LHsType GhcPs)

Decompose a GADT type into its constituent parts. Returns (outer_bndrs, mb_ctxt, body), where:

outer_bndrs are HsOuterExplicit if the type has explicit, outermost type variable binders. Otherwise, they are HsOuterImplicit.

mb_ctxt is Just the context, if it is provided. Otherwise, it is Nothing.

body is the body of the type after the optional foralls and context.

This function is careful not to look through parentheses. See Note [GADT abstract syntax] (Wrinkle: No nested foralls or contexts) GHC.Hs.Decls for why this is important.

splitHsFunType :: forall (p :: Pass). LHsType (GhcPass p) -> (([EpToken "("], [EpToken ")"]), EpAnnComments, [HsScaled (GhcPass p) (LHsType (GhcPass p))], LHsType (GhcPass p))

hsTyGetAppHead_maybe :: forall (p :: Pass). Anno (IdGhcP p) ~ SrcSpanAnnN => LHsType (GhcPass p) -> Maybe (LocatedN (IdP (GhcPass p)))

Retrieve the name of the "head" of a nested type application. This is somewhat like GHC.Tc.Gen.HsType.splitHsAppTys, but a little more thorough. The purpose of this function is to examine instance heads, so it doesn't handle *all* cases (like lists, tuples, (~), etc.).

mkHsOpTy :: forall (p :: Pass). Anno (IdGhcP p) ~ SrcSpanAnnN => PromotionFlag -> LHsType (GhcPass p) -> LocatedN (IdP (GhcPass p)) -> LHsType (GhcPass p) -> HsType (GhcPass p)

mkHsAppTy :: forall (p :: Pass). LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

mkHsAppTys :: forall (p :: Pass). LHsType (GhcPass p) -> [LHsType (GhcPass p)] -> LHsType (GhcPass p)

mkHsAppKindTy :: forall (p :: Pass). XAppKindTy (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

ignoreParens :: forall (p :: Pass). LHsType (GhcPass p) -> LHsType (GhcPass p)

hsSigWcType :: UnXRec p => LHsSigWcType p -> LHsType p

hsPatSigType :: HsPatSigType pass -> LHsType pass

hsTyKindSig :: forall (p :: Pass). LHsType (GhcPass p) -> Maybe (LHsKind (GhcPass p))

Get the kind signature of a type, ignoring parentheses:

hsTyKindSig `Maybe ` = Nothing hsTyKindSig `Maybe :: Type -> Type ` = Just `Type -> Type` hsTyKindSig `Maybe :: ((Type -> Type))` = Just `Type -> Type`

This is used to extract the result kind of type synonyms with a CUSK:

type S = (F :: res_kind) ^^^^^^^^

setHsTyVarBndrFlag :: forall flag flag' (pass :: Pass). flag -> HsTyVarBndr flag' (GhcPass pass) -> HsTyVarBndr flag (GhcPass pass)

Set the attached flag

hsTyVarBndrFlag :: forall flag (pass :: Pass). HsTyVarBndr flag (GhcPass pass) -> flag

Return the attached flag

updateHsTyVarBndrFlag :: forall flag flag' (pass :: Pass). (flag -> flag') -> HsTyVarBndr flag (GhcPass pass) -> HsTyVarBndr flag' (GhcPass pass)

Update the attached flag

pprHsType :: forall (p :: Pass). OutputableBndrId p => HsType (GhcPass p) -> SDoc

pprHsForAll :: forall (p :: Pass). OutputableBndrId p => HsForAllTelescope (GhcPass p) -> Maybe (LHsContext (GhcPass p)) -> SDoc

Prints a forall; When passed an empty list, prints forall ./forall -> only when -dppr-debug is enabled.

pprHsOuterFamEqnTyVarBndrs :: forall (p :: Pass). OutputableBndrId p => HsOuterFamEqnTyVarBndrs (GhcPass p) -> SDoc

Prints the explicit forall in a type family equation if one is written. If there is no explicit forall, nothing is printed.

pprHsOuterSigTyVarBndrs :: forall (p :: Pass). OutputableBndrId p => HsOuterSigTyVarBndrs (GhcPass p) -> SDoc

Prints the outermost forall in a type signature if one is written. If there is no outermost forall, nothing is printed.

pprLHsContext :: forall (p :: Pass). OutputableBndrId p => Maybe (LHsContext (GhcPass p)) -> SDoc

hsTypeNeedsParens :: forall (p :: Pass). PprPrec -> HsType (GhcPass p) -> Bool

hsTypeNeedsParens p t returns True if the type t needs parentheses under precedence p.

parenthesizeHsType :: forall (p :: Pass). PprPrec -> LHsType (GhcPass p) -> LHsType (GhcPass p)

parenthesizeHsType p ty checks if hsTypeNeedsParens p ty is true, and if so, surrounds ty with an HsParTy. Otherwise, it simply returns ty.

parenthesizeHsContext :: forall (p :: Pass). PprPrec -> LHsContext (GhcPass p) -> LHsContext (GhcPass p)

parenthesizeHsContext p ctxt checks if ctxt is a single constraint c such that hsTypeNeedsParens p c is true, and if so, surrounds c with an HsParTy to form a parenthesized ctxt. Otherwise, it simply returns ctxt unchanged.

Orphan instances

Outputable HsIPName

Instance details

Methods

ppr :: HsIPName -> SDoc

OutputableBndr HsIPName

Instance details

Methods

pprBndr :: BindingSite -> HsIPName -> SDoc

pprPrefixOcc :: HsIPName -> SDoc

pprInfixOcc :: HsIPName -> SDoc

bndrIsJoin_maybe :: HsIPName -> JoinPointHood

OutputableBndrId p => Outputable (ConDeclField (GhcPass p))

Instance details

Methods

ppr :: ConDeclField (GhcPass p) -> SDoc

Outputable (XRec pass (IdP pass)) => Outputable (FieldOcc pass)

Instance details

Methods

ppr :: FieldOcc pass -> SDoc

OutputableBndrId p => Outputable (HsBndrVar (GhcPass p))

Instance details

Methods

ppr :: HsBndrVar (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsForAllTelescope (GhcPass p))

Instance details

Methods

ppr :: HsForAllTelescope (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsPatSigType (GhcPass p))

Instance details

Methods

ppr :: HsPatSigType (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsSigType (GhcPass p))

Instance details

Methods

ppr :: HsSigType (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsTyLit (GhcPass p))

Instance details

Methods

ppr :: HsTyLit (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsTyPat (GhcPass p))

Instance details

Methods

ppr :: HsTyPat (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsType (GhcPass p))

Instance details

Methods

ppr :: HsType (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (LHsQTyVars (GhcPass p))

Instance details

Methods

ppr :: LHsQTyVars (GhcPass p) -> SDoc

OutputableBndrId pass => OutputableBndr (FieldOcc (GhcPass pass))

Instance details

Methods

pprBndr :: BindingSite -> FieldOcc (GhcPass pass) -> SDoc

pprPrefixOcc :: FieldOcc (GhcPass pass) -> SDoc

pprInfixOcc :: FieldOcc (GhcPass pass) -> SDoc

bndrIsJoin_maybe :: FieldOcc (GhcPass pass) -> JoinPointHood

(Outputable mult, OutputableBndrId pass) => Outputable (HsArrowOf mult (GhcPass pass))

Instance details

Methods

ppr :: HsArrowOf mult (GhcPass pass) -> SDoc

(OutputableBndrFlag flag p, OutputableBndrFlag flag (NoGhcTcPass p), OutputableBndrId p) => Outputable (HsOuterTyVarBndrs flag (GhcPass p))

Instance details

Methods

ppr :: HsOuterTyVarBndrs flag (GhcPass p) -> SDoc

(OutputableBndrId p, OutputableBndrFlag flag p) => Outputable (HsTyVarBndr flag (GhcPass p))

Instance details

Methods

ppr :: HsTyVarBndr flag (GhcPass p) -> SDoc

Outputable thing => Outputable (HsWildCardBndrs (GhcPass p) thing)

Instance details

Methods

ppr :: HsWildCardBndrs (GhcPass p) thing -> SDoc

OutputableBndrId pass => OutputableBndr (GenLocated SrcSpan (FieldOcc (GhcPass pass)))

Instance details

Methods

pprBndr :: BindingSite -> GenLocated SrcSpan (FieldOcc (GhcPass pass)) -> SDoc

pprPrefixOcc :: GenLocated SrcSpan (FieldOcc (GhcPass pass)) -> SDoc

pprInfixOcc :: GenLocated SrcSpan (FieldOcc (GhcPass pass)) -> SDoc

bndrIsJoin_maybe :: GenLocated SrcSpan (FieldOcc (GhcPass pass)) -> JoinPointHood

(Outputable tm, Outputable ty) => Outputable (HsArg (GhcPass p) tm ty)
This instance is meant for debug-printing purposes. If you wish to pretty-print an application of HsArgs, use pprHsArgsApp instead.

Instance details

Methods

ppr :: HsArg (GhcPass p) tm ty -> SDoc

(Outputable tyarg, Outputable arg, Outputable rec) => Outputable (HsConDetails tyarg arg rec)

Instance details

Methods

ppr :: HsConDetails tyarg arg rec -> SDoc