GHC.Hs.Type

Documentation

type Mult = Type

Mult is a type alias for Type.

Mult must contain Type because multiplicity variables are mere type variables (of kind Multiplicity) in Haskell. So the simplest implementation is to make Mult be Type.

Multiplicities can be formed with: - One: GHC.Types.One (= oneDataCon) - Many: GHC.Types.Many (= manyDataCon) - Multiplication: GHC.Types.MultMul (= multMulTyCon)

So that Mult feels a bit more structured, we provide pattern synonyms and smart constructors for these.

data HsScaled pass a

This is used in the syntax. In constructor declaration. It must keep the arrow representation.

Constructors

HsScaled (HsArrow pass) a

hsMult :: HsScaled pass a -> HsArrow pass

hsScaledThing :: HsScaled pass a -> a

data HsArrow pass

Denotes the type of arrows in the surface language

Constructors

HsUnrestrictedArrow IsUnicodeSyntax
a -> b or a → b

HsLinearArrow IsUnicodeSyntax (Maybe AddEpAnn)
a %1 -> b or a %1 → b, or a ⊸ b

HsExplicitMult IsUnicodeSyntax (Maybe AddEpAnn) (LHsType pass)
a %m -> b or a %m → b (very much including `a %Many -> b`! This is how the programmer wrote it). It is stored as an HsType so as to preserve the syntax as written in the program.

arrowToHsType :: HsArrow GhcRn -> LHsType GhcRn

Convert an arrow into its corresponding multiplicity. In essence this erases the information of whether the programmer wrote an explicit multiplicity or a shorthand.

hsLinear :: a -> HsScaled pass a

When creating syntax we use the shorthands. It's better for printing, also, the shorthands work trivially at each pass.

hsUnrestricted :: a -> HsScaled pass a

When creating syntax we use the shorthands. It's better for printing, also, the shorthands work trivially at each pass.

isUnrestricted :: HsArrow GhcRn -> Bool

data HsType pass

Haskell Type

Constructors

HsForAllTy

AnnKeywordId : AnnForall, AnnDot,AnnDarrow For details on above see note [exact print annotations] in GHC.Parser.Annotation

Fields

hst_xforall :: XForAllTy pass

hst_tele :: HsForAllTelescope pass

hst_body :: LHsType pass

HsQualTy

Fields

hst_xqual :: XQualTy pass

hst_ctxt :: Maybe (LHsContext pass)

hst_body :: LHsType pass

HsTyVar (XTyVar pass) PromotionFlag (LIdP pass)

AnnKeywordId : None

HsAppTy (XAppTy pass) (LHsType pass) (LHsType pass)

AnnKeywordId : None

HsAppKindTy (XAppKindTy pass) (LHsType pass) (LHsKind pass)

HsFunTy (XFunTy pass) (HsArrow pass) (LHsType pass) (LHsType pass)

AnnKeywordId : AnnRarrow,

HsListTy (XListTy pass) (LHsType pass)

AnnKeywordId : AnnOpen '[', AnnClose ']'

HsTupleTy (XTupleTy pass) HsTupleSort [LHsType pass]

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

HsSumTy (XSumTy pass) [LHsType pass]

AnnKeywordId : AnnOpen '(#', AnnClose '#)'@

HsOpTy (XOpTy pass) (LHsType pass) (LIdP pass) (LHsType pass)

AnnKeywordId : None

HsParTy (XParTy pass) (LHsType pass)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsIParamTy (XIParamTy pass) (XRec pass HsIPName) (LHsType pass)
(?x :: ty)

AnnKeywordId : AnnDcolon

HsStarTy (XStarTy pass) Bool

AnnKeywordId : None

HsKindSig (XKindSig pass) (LHsType pass) (LHsKind pass)
(ty :: kind)

AnnKeywordId : AnnOpen '(', AnnDcolon,AnnClose ')'

HsSpliceTy (XSpliceTy pass) (HsSplice pass)

AnnKeywordId : AnnOpen '$(', AnnClose ')'

HsDocTy (XDocTy pass) (LHsType pass) LHsDocString

AnnKeywordId : None

HsBangTy (XBangTy pass) HsSrcBang (LHsType pass)

AnnKeywordId : AnnOpen '{-# UNPACK' or '{-# NOUNPACK', AnnClose '#-}' AnnBang '!'

HsRecTy (XRecTy pass) [LConDeclField pass]

AnnKeywordId : AnnOpen '{', AnnClose '}'

HsExplicitListTy (XExplicitListTy pass) PromotionFlag [LHsType pass]

AnnKeywordId : AnnOpen "'[", AnnClose ']'

HsExplicitTupleTy (XExplicitTupleTy pass) [LHsType pass]

AnnKeywordId : AnnOpen "'(", AnnClose ')'

HsTyLit (XTyLit pass) HsTyLit

AnnKeywordId : None

HsWildCardTy (XWildCardTy pass)

AnnKeywordId : None

XHsType !(XXType pass)

type HsCoreTy = Type

type LHsType pass

Arguments

= XRec pass (HsType pass)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Type

type HsKind pass = HsType pass

Haskell Kind

type LHsKind pass

Arguments

= XRec pass (HsKind pass)
AnnKeywordId : AnnDcolon

Located Haskell Kind

data HsForAllTelescope pass

The type variable binders in an HsForAllTy. See also Note [Variable Specificity and Forall Visibility] in GHC.Tc.Gen.HsType.

Constructors

HsForAllVis
A visible forall (e.g., forall a -> {...}). These do not have any notion of specificity, so we use () as a placeholder value.

Fields

hsf_xvis :: XHsForAllVis pass

hsf_vis_bndrs :: [LHsTyVarBndr () pass]

HsForAllInvis
An invisible forall (e.g., forall a {b} c. {...}), where each binder has a Specificity.

Fields

hsf_xinvis :: XHsForAllInvis pass

hsf_invis_bndrs :: [LHsTyVarBndr Specificity pass]

XHsForAllTelescope !(XXHsForAllTelescope pass)

type EpAnnForallTy

Arguments

= EpAnn (AddEpAnn, AddEpAnn)
Location of forall and -> for HsForAllVis Location of forall and . for HsForAllInvis

data HsTyVarBndr flag pass

Haskell Type Variable Binder The flag annotates the binder. It is Specificity in places where explicit specificity is allowed (e.g. x :: forall {a} b. ...) or () in other places.

Constructors

UserTyVar (XUserTyVar pass) flag (LIdP pass)

KindedTyVar (XKindedTyVar pass) flag (LIdP pass) (LHsKind pass)

AnnKeywordId : AnnOpen, AnnDcolon, AnnClose

XTyVarBndr !(XXTyVarBndr pass)

type LHsTyVarBndr flag pass = XRec pass (HsTyVarBndr flag pass)

Located Haskell Type Variable Binder

data LHsQTyVars pass

Located Haskell Quantified Type Variables

Constructors

HsQTvs

Fields

hsq_ext :: XHsQTvs pass

hsq_explicit :: [LHsTyVarBndr () pass]

XLHsQTyVars !(XXLHsQTyVars pass)

data HsOuterTyVarBndrs flag pass

The outermost type variables in a type that obeys the forall-or-nothing rule. See Note [forall-or-nothing rule].

Constructors

HsOuterImplicit
Implicit forall, e.g., f :: a -> b -> b

Fields

hso_ximplicit :: XHsOuterImplicit pass

HsOuterExplicit
Explicit forall, e.g., f :: forall a b. a -> b -> b

Fields

hso_xexplicit :: XHsOuterExplicit pass flag

hso_bndrs :: [LHsTyVarBndr flag (NoGhcTc pass)]

XHsOuterTyVarBndrs !(XXHsOuterTyVarBndrs pass)

type HsOuterFamEqnTyVarBndrs = HsOuterTyVarBndrs ()

Used for type-family instance equations, e.g.,

type instance forall a. F [a] = Tree a

The notion of specificity is irrelevant in type family equations, so we use () for the HsOuterTyVarBndrs flag.

type HsOuterSigTyVarBndrs = HsOuterTyVarBndrs Specificity

Used for signatures, e.g.,

f :: forall a {b}. blah

We use Specificity for the HsOuterTyVarBndrs flag to allow distinguishing between specified and inferred type variables.

data HsWildCardBndrs pass thing

Haskell Wildcard Binders

Constructors

HsWC

Fields

hswc_ext :: XHsWC pass thing

hswc_body :: thing

XHsWildCardBndrs !(XXHsWildCardBndrs pass thing)

data HsPatSigType pass

Types that can appear in pattern signatures, as well as the signatures for term-level binders in RULES. See Note [Pattern signature binders and scoping].

This is very similar to HsSigWcType, but with slightly different semantics: see Note [HsType binders]. See also Note [The wildcard story for types].

Constructors

HsPS

Fields

hsps_ext :: XHsPS pass

After renamer: HsPSRn

hsps_body :: LHsType pass

Main payload (the type itself)

XHsPatSigType !(XXHsPatSigType pass)

data HsPSRn

The extension field for HsPatSigType, which is only used in the renamer onwards. See Note [Pattern signature binders and scoping].

Constructors

HsPSRn

Fields

hsps_nwcs :: [Name]

Wildcard names

hsps_imp_tvs :: [Name]

Implicitly bound variable names

data HsSigType pass

A type signature that obeys the forall-or-nothing rule. In other words, an LHsType that uses an HsOuterSigTyVarBndrs to represent its outermost type variable quantification. See Note [Representing type signatures].

Constructors

HsSig

Fields

sig_ext :: XHsSig pass

sig_bndrs :: HsOuterSigTyVarBndrs pass

sig_body :: LHsType pass

XHsSigType !(XXHsSigType pass)

type LHsSigType pass = XRec pass (HsSigType pass)

Located Haskell Signature Type

type LHsSigWcType pass = HsWildCardBndrs pass (LHsSigType pass)

Located Haskell Signature Wildcard Type

type LHsWcType pass = HsWildCardBndrs pass (LHsType pass)

Located Haskell Wildcard Type

data HsTupleSort

Haskell Tuple Sort

Constructors

HsUnboxedTuple

HsBoxedOrConstraintTuple

type HsContext pass = [LHsType pass]

Haskell Context

type LHsContext pass

Arguments

= XRec pass (HsContext pass)
AnnKeywordId : AnnUnit For details on above see note [exact print annotations] in GHC.Parser.Annotation

Located Haskell Context

fromMaybeContext :: Maybe (LHsContext (GhcPass p)) -> HsContext (GhcPass p)

data HsTyLit

Haskell Type Literal

Constructors

HsNumTy SourceText Integer

HsStrTy SourceText FastString

HsCharTy SourceText Char

newtype HsIPName

These names are used early on to store the names of implicit parameters. They completely disappear after type-checking.

Constructors

HsIPName FastString

hsIPNameFS :: HsIPName -> FastString

data HsArg tm ty

Constructors

HsValArg tm

HsTypeArg SrcSpan ty

HsArgPar SrcSpan

numVisibleArgs :: [HsArg tm ty] -> Arity

type LHsTypeArg p = HsArg (LHsType p) (LHsKind p)

lhsTypeArgSrcSpan :: LHsTypeArg (GhcPass pass) -> SrcSpan

Compute the SrcSpan associated with an LHsTypeArg.

class OutputableBndrFlag flag p

Minimal complete definition

pprTyVarBndr

type LBangType pass = XRec pass (BangType pass)

Located Bang Type

type BangType pass = HsType pass

Bang Type

In the parser, strictness and packedness annotations bind more tightly than docstrings. This means that when consuming a BangType (and looking for HsBangTy) we must be ready to peer behind a potential layer of HsDocTy. See #15206 for motivation and getBangType for an example.

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as the user wrote them in the source code.

(HsSrcBang _ SrcUnpack SrcLazy) and (HsSrcBang _ SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsSrcBang _ NoSrcUnpack SrcLazy)

Constructors

HsSrcBang SourceText SrcUnpackedness SrcStrictness

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field, or one with an unlifted type

HsStrict
Strict but not unpacked field

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie ~

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{-# UNPACK #-} specified

SrcNoUnpack
{-# NOUNPACK #-} specified

NoSrcUnpack
no unpack pragma

getBangType :: LHsType (GhcPass p) -> LHsType (GhcPass p)

getBangStrictness :: LHsType (GhcPass p) -> HsSrcBang

data ConDeclField pass

Constructor Declaration Field

Constructors

ConDeclField

Fields

cd_fld_ext :: XConDeclField pass

cd_fld_names :: [LFieldOcc pass]

See Note [ConDeclField passs]

cd_fld_type :: LBangType pass

cd_fld_doc :: Maybe LHsDocString

AnnKeywordId : AnnDcolon

XConDeclField !(XXConDeclField pass)

type LConDeclField pass

Arguments

= XRec pass (ConDeclField pass)
May have AnnKeywordId : AnnComma when in a list

Located Constructor Declaration Field

pprConDeclFields :: OutputableBndrId p => [LConDeclField (GhcPass p)] -> SDoc

data HsConDetails tyarg arg rec

Describes the arguments to a data constructor. This is a common representation for several constructor-related concepts, including:

The arguments in a Haskell98-style constructor declaration (see HsConDeclH98Details in GHC.Hs.Decls).

The arguments in constructor patterns in case/function definitions (see HsConPatDetails in GHC.Hs.Pat).

The left-hand side arguments in a pattern synonym binding (see HsPatSynDetails in GHC.Hs.Binds).

One notable exception is the arguments in a GADT constructor, which uses a separate data type entirely (see HsConDeclGADTDetails in GHC.Hs.Decls). This is because GADT constructors cannot be declared with infix syntax, unlike the concepts above (#18844).

Constructors

PrefixCon [tyarg] [arg]

RecCon rec

InfixCon arg arg

noTypeArgs :: [Void]

An empty list that can be used to indicate that there are no type arguments allowed in cases where HsConDetails is applied to Void.

data FieldOcc pass

Field Occurrence

Represents an *occurrence* of an unambiguous field. This may or may not be a binding occurrence (e.g. this type is used in ConDeclField and RecordPatSynField which bind their fields, but also in HsRecField for record construction and patterns, which do not).

We store both the RdrName the user originally wrote, and after the renamer, the selector function.

Constructors

FieldOcc

Fields

extFieldOcc :: XCFieldOcc pass

rdrNameFieldOcc :: LocatedN RdrName

See Note [Located RdrNames] in GHC.Hs.Expr

XFieldOcc !(XXFieldOcc pass)

type LFieldOcc pass = XRec pass (FieldOcc pass)

Located Field Occurrence

mkFieldOcc :: LocatedN RdrName -> FieldOcc GhcPs

data AmbiguousFieldOcc pass

Ambiguous Field Occurrence

Represents an *occurrence* of a field that is potentially ambiguous after the renamer, with the ambiguity resolved by the typechecker. We always store the RdrName that the user originally wrote, and store the selector function after the renamer (for unambiguous occurrences) or the typechecker (for ambiguous occurrences).

See Note [HsRecField and HsRecUpdField] in GHC.Hs.Pat and Note [Disambiguating record fields] in GHC.Tc.Gen.Head. See Note [Located RdrNames] in GHC.Hs.Expr

Constructors

Unambiguous (XUnambiguous pass) (LocatedN RdrName)

Ambiguous (XAmbiguous pass) (LocatedN RdrName)

XAmbiguousFieldOcc !(XXAmbiguousFieldOcc pass)

mkAmbiguousFieldOcc :: LocatedN RdrName -> AmbiguousFieldOcc GhcPs

rdrNameAmbiguousFieldOcc :: AmbiguousFieldOcc (GhcPass p) -> RdrName

selectorAmbiguousFieldOcc :: AmbiguousFieldOcc GhcTc -> Id

unambiguousFieldOcc :: AmbiguousFieldOcc GhcTc -> FieldOcc GhcTc

ambiguousFieldOcc :: FieldOcc GhcTc -> AmbiguousFieldOcc GhcTc

mkAnonWildCardTy :: HsType GhcPs

pprAnonWildCard :: SDoc

hsOuterTyVarNames :: HsOuterTyVarBndrs flag GhcRn -> [Name]

hsOuterExplicitBndrs :: HsOuterTyVarBndrs flag (GhcPass p) -> [LHsTyVarBndr flag (NoGhcTc (GhcPass p))]

mapHsOuterImplicit :: (XHsOuterImplicit pass -> XHsOuterImplicit pass) -> HsOuterTyVarBndrs flag pass -> HsOuterTyVarBndrs flag pass

mkHsOuterImplicit :: HsOuterTyVarBndrs flag GhcPs

mkHsOuterExplicit :: EpAnnForallTy -> [LHsTyVarBndr flag GhcPs] -> HsOuterTyVarBndrs flag GhcPs

mkHsImplicitSigType :: LHsType GhcPs -> HsSigType GhcPs

mkHsExplicitSigType :: EpAnnForallTy -> [LHsTyVarBndr Specificity GhcPs] -> LHsType GhcPs -> HsSigType GhcPs

mkHsWildCardBndrs :: thing -> HsWildCardBndrs GhcPs thing

mkHsPatSigType :: EpAnn EpaLocation -> LHsType GhcPs -> HsPatSigType GhcPs

mkEmptyWildCardBndrs :: thing -> HsWildCardBndrs GhcRn thing

mkHsForAllVisTele :: EpAnnForallTy -> [LHsTyVarBndr () (GhcPass p)] -> HsForAllTelescope (GhcPass p)

mkHsForAllInvisTele :: EpAnnForallTy -> [LHsTyVarBndr Specificity (GhcPass p)] -> HsForAllTelescope (GhcPass p)

mkHsQTvs :: [LHsTyVarBndr () GhcPs] -> LHsQTyVars GhcPs

hsQTvExplicit :: LHsQTyVars pass -> [LHsTyVarBndr () pass]

emptyLHsQTvs :: LHsQTyVars GhcRn

isHsKindedTyVar :: HsTyVarBndr flag pass -> Bool

Does this HsTyVarBndr come with an explicit kind annotation?

hsTvbAllKinded :: LHsQTyVars (GhcPass p) -> Bool

Do all type variables in this LHsQTyVars come with kind annotations?

hsScopedTvs :: LHsSigType GhcRn -> [Name]

hsWcScopedTvs :: LHsSigWcType GhcRn -> [Name]

dropWildCards :: LHsSigWcType pass -> LHsSigType pass

hsTyVarName :: HsTyVarBndr flag (GhcPass p) -> IdP (GhcPass p)

hsAllLTyVarNames :: LHsQTyVars GhcRn -> [Name]

hsLTyVarLocNames :: LHsQTyVars (GhcPass p) -> [LocatedN (IdP (GhcPass p))]

hsLTyVarName :: LHsTyVarBndr flag (GhcPass p) -> IdP (GhcPass p)

hsLTyVarNames :: [LHsTyVarBndr flag (GhcPass p)] -> [IdP (GhcPass p)]

hsLTyVarLocName :: LHsTyVarBndr flag (GhcPass p) -> LocatedN (IdP (GhcPass p))

hsExplicitLTyVarNames :: LHsQTyVars (GhcPass p) -> [IdP (GhcPass p)]

splitLHsInstDeclTy :: LHsSigType GhcRn -> ([Name], Maybe (LHsContext GhcRn), LHsType GhcRn)

Decompose a type class instance type (of the form forall tvs. context => instance_head) into its constituent parts. Note that the [Name]s returned correspond to either:

The implicitly bound type variables (if the type lacks an outermost forall), or

The explicitly bound type variables (if the type has an outermost forall).

This function is careful not to look through parentheses. See Note [No nested foralls or contexts in instance types] for why this is important.

getLHsInstDeclHead :: LHsSigType (GhcPass p) -> LHsType (GhcPass p)

Decompose a type class instance type (of the form forall tvs. context => instance_head) into the instance_head.

getLHsInstDeclClass_maybe :: Anno (IdGhcP p) ~ SrcSpanAnnN => LHsSigType (GhcPass p) -> Maybe (LocatedN (IdP (GhcPass p)))

Decompose a type class instance type (of the form forall tvs. context => instance_head) into the instance_head and retrieve the underlying class type constructor (if it exists).

splitLHsPatSynTy :: LHsSigType (GhcPass p) -> ([LHsTyVarBndr Specificity (GhcPass (NoGhcTcPass p))], Maybe (LHsContext (GhcPass p)), [LHsTyVarBndr Specificity (GhcPass p)], Maybe (LHsContext (GhcPass p)), LHsType (GhcPass p))

Decompose a pattern synonym type signature into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsForAllTyInvis :: LHsType (GhcPass pass) -> ((EpAnnForallTy, [LHsTyVarBndr Specificity (GhcPass pass)]), LHsType (GhcPass pass))

Decompose a type of the form forall tvs. body into its constituent parts. Only splits type variable binders that were quantified invisibly (e.g., forall a., with a dot).

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them. Unlike splitLHsSigmaTyInvis, this function does not look through parentheses, hence the suffix _KP (short for "Keep Parentheses").

splitLHsForAllTyInvis_KP :: LHsType (GhcPass pass) -> (Maybe (EpAnnForallTy, [LHsTyVarBndr Specificity (GhcPass pass)]), LHsType (GhcPass pass))

Decompose a type of the form forall tvs. body into its constituent parts. Only splits type variable binders that were quantified invisibly (e.g., forall a., with a dot).

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Unlike splitLHsForAllTyInvis, this function does not look through parentheses, hence the suffix _KP (short for "Keep Parentheses").

splitLHsQualTy :: LHsType (GhcPass pass) -> (Maybe (LHsContext (GhcPass pass)), LHsType (GhcPass pass))

Decompose a type of the form context => body into its constituent parts.

Note that this function looks through parentheses, so it will work on types such as (context => ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsSigmaTyInvis :: LHsType (GhcPass p) -> ([LHsTyVarBndr Specificity (GhcPass p)], Maybe (LHsContext (GhcPass p)), LHsType (GhcPass p))

Decompose a sigma type (of the form forall tvs. context => body) into its constituent parts. Only splits type variable binders that were quantified invisibly (e.g., forall a., with a dot).

This function is used to split apart certain types, such as instance declaration types, which disallow visible foralls. For instance, if GHC split apart the forall in instance forall a -> Show (Blah a), then that declaration would mistakenly be accepted!

Note that this function looks through parentheses, so it will work on types such as (forall a. ...). The downside to this is that it is not generally possible to take the returned types and reconstruct the original type (parentheses and all) from them.

splitLHsGadtTy :: LHsSigType GhcPs -> (HsOuterSigTyVarBndrs GhcPs, Maybe (LHsContext GhcPs), LHsType GhcPs)

Decompose a GADT type into its constituent parts. Returns (outer_bndrs, mb_ctxt, body), where:

outer_bndrs are HsOuterExplicit if the type has explicit, outermost type variable binders. Otherwise, they are HsOuterImplicit.

mb_ctxt is Just the context, if it is provided. Otherwise, it is Nothing.

body is the body of the type after the optional foralls and context.

This function is careful not to look through parentheses. See Note [GADT abstract syntax] (Wrinkle: No nested foralls or contexts) GHC.Hs.Decls for why this is important.

splitHsFunType :: LHsType (GhcPass p) -> ([AddEpAnn], EpAnnComments, [HsScaled (GhcPass p) (LHsType (GhcPass p))], LHsType (GhcPass p))

hsTyGetAppHead_maybe :: Anno (IdGhcP p) ~ SrcSpanAnnN => LHsType (GhcPass p) -> Maybe (LocatedN (IdP (GhcPass p)))

Retrieve the name of the "head" of a nested type application. This is somewhat like GHC.Tc.Gen.HsType.splitHsAppTys, but a little more thorough. The purpose of this function is to examine instance heads, so it doesn't handle *all* cases (like lists, tuples, (~), etc.).

mkHsOpTy :: Anno (IdGhcP p) ~ SrcSpanAnnN => LHsType (GhcPass p) -> LocatedN (IdP (GhcPass p)) -> LHsType (GhcPass p) -> HsType (GhcPass p)

mkHsAppTy :: LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

mkHsAppTys :: LHsType (GhcPass p) -> [LHsType (GhcPass p)] -> LHsType (GhcPass p)

mkHsAppKindTy :: XAppKindTy (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p) -> LHsType (GhcPass p)

ignoreParens :: LHsType (GhcPass p) -> LHsType (GhcPass p)

hsSigWcType :: forall p. UnXRec p => LHsSigWcType p -> LHsType p

hsPatSigType :: HsPatSigType pass -> LHsType pass

hsTyKindSig :: LHsType (GhcPass p) -> Maybe (LHsKind (GhcPass p))

Get the kind signature of a type, ignoring parentheses:

hsTyKindSig `Maybe ` = Nothing hsTyKindSig `Maybe :: Type -> Type ` = Just `Type -> Type` hsTyKindSig `Maybe :: ((Type -> Type))` = Just `Type -> Type`

This is used to extract the result kind of type synonyms with a CUSK:

type S = (F :: res_kind) ^^^^^^^^

setHsTyVarBndrFlag :: flag -> HsTyVarBndr flag' (GhcPass pass) -> HsTyVarBndr flag (GhcPass pass)

Set the attached flag

hsTyVarBndrFlag :: HsTyVarBndr flag (GhcPass pass) -> flag

Return the attached flag

pprHsType :: OutputableBndrId p => HsType (GhcPass p) -> SDoc

pprHsForAll :: forall p. OutputableBndrId p => HsForAllTelescope (GhcPass p) -> Maybe (LHsContext (GhcPass p)) -> SDoc

Prints a forall; When passed an empty list, prints forall ./forall -> only when -dppr-debug is enabled.

pprHsOuterFamEqnTyVarBndrs :: OutputableBndrId p => HsOuterFamEqnTyVarBndrs (GhcPass p) -> SDoc

Prints the explicit forall in a type family equation if one is written. If there is no explicit forall, nothing is printed.

pprHsOuterSigTyVarBndrs :: OutputableBndrId p => HsOuterSigTyVarBndrs (GhcPass p) -> SDoc

Prints the outermost forall in a type signature if one is written. If there is no outermost forall, nothing is printed.

pprLHsContext :: OutputableBndrId p => Maybe (LHsContext (GhcPass p)) -> SDoc

hsTypeNeedsParens :: PprPrec -> HsType (GhcPass p) -> Bool

hsTypeNeedsParens p t returns True if the type t needs parentheses under precedence p.

parenthesizeHsType :: PprPrec -> LHsType (GhcPass p) -> LHsType (GhcPass p)

parenthesizeHsType p ty checks if hsTypeNeedsParens p ty is true, and if so, surrounds ty with an HsParTy. Otherwise, it simply returns ty.

parenthesizeHsContext :: PprPrec -> LHsContext (GhcPass p) -> LHsContext (GhcPass p)

parenthesizeHsContext p ctxt checks if ctxt is a single constraint c such that hsTypeNeedsParens p c is true, and if so, surrounds c with an HsParTy to form a parenthesized ctxt. Otherwise, it simply returns ctxt unchanged.

Orphan instances

Outputable (AmbiguousFieldOcc (GhcPass p))

Instance details

Methods

ppr :: AmbiguousFieldOcc (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (ConDeclField (GhcPass p))

Instance details

Methods

ppr :: ConDeclField (GhcPass p) -> SDoc

OutputableBndrId pass => Outputable (HsArrow (GhcPass pass))

Instance details

Methods

ppr :: HsArrow (GhcPass pass) -> SDoc

OutputableBndrId p => Outputable (HsForAllTelescope (GhcPass p))

Instance details

Methods

ppr :: HsForAllTelescope (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsPatSigType (GhcPass p))

Instance details

Methods

ppr :: HsPatSigType (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsSigType (GhcPass p))

Instance details

Methods

ppr :: HsSigType (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (HsType (GhcPass p))

Instance details

Methods

ppr :: HsType (GhcPass p) -> SDoc

OutputableBndrId p => Outputable (LHsQTyVars (GhcPass p))

Instance details

Methods

ppr :: LHsQTyVars (GhcPass p) -> SDoc

OutputableBndr (AmbiguousFieldOcc (GhcPass p))

Instance details

Methods

pprBndr :: BindingSite -> AmbiguousFieldOcc (GhcPass p) -> SDoc

pprPrefixOcc :: AmbiguousFieldOcc (GhcPass p) -> SDoc

pprInfixOcc :: AmbiguousFieldOcc (GhcPass p) -> SDoc

bndrIsJoin_maybe :: AmbiguousFieldOcc (GhcPass p) -> Maybe Int

NamedThing (HsTyVarBndr flag GhcRn)

Instance details

Methods

getOccName :: HsTyVarBndr flag GhcRn -> OccName

getName :: HsTyVarBndr flag GhcRn -> Name

(OutputableBndrFlag flag p, OutputableBndrFlag flag (NoGhcTcPass p), OutputableBndrId p) => Outputable (HsOuterTyVarBndrs flag (GhcPass p))

Instance details

Methods

ppr :: HsOuterTyVarBndrs flag (GhcPass p) -> SDoc

(OutputableBndrId p, OutputableBndrFlag flag p) => Outputable (HsTyVarBndr flag (GhcPass p))

Instance details

Methods

ppr :: HsTyVarBndr flag (GhcPass p) -> SDoc

Outputable thing => Outputable (HsWildCardBndrs (GhcPass p) thing)

Instance details

Methods

ppr :: HsWildCardBndrs (GhcPass p) thing -> SDoc