GHC.HsToCore.Pmc.Types

Description

Types used through-out pattern match checking. This module is mostly there to be imported from GHC.HsToCore.Types. The exposed API is that of GHC.HsToCore.Pmc.

These types model the paper Lower Your Guards: A Compositional Pattern-Match Coverage Checker".

LYG syntax

Guard language

newtype SrcInfo

Means by which we identify a source construct for later pretty-printing in a warning message. SDoc for the equation to show, Located for the location.

Constructors

SrcInfo (Located SDoc)

data PmGrd

A very simple language for pattern guards. Let bindings, bang patterns, and matching variables against flat constructor patterns. The LYG guard language.

Constructors

PmCon
PmCon x K dicts args corresponds to a K dicts args <- x guard. The args are bound in this construct, the x is just a use. For the arguments' meaning see ConPatOut.

Fields

pm_id :: !Id

pm_con_con :: !PmAltCon

pm_con_tvs :: ![TyVar]

pm_con_dicts :: ![EvVar]

pm_con_args :: ![Id]

PmBang
PmBang x corresponds to a seq x True guard. If the extra SrcInfo is present, the bang guard came from a source bang pattern, in which case we might want to report it as redundant. See Note [Dead bang patterns] in GHC.HsToCore.Pmc.Check.

Fields

pm_id :: !Id

_pm_loc :: !(Maybe SrcInfo)

PmLet
PmLet x expr corresponds to a let x = expr guard. This actually binds x.

Fields

pm_id :: !Id

_pm_let_expr :: !CoreExpr

newtype GrdVec

A sequence of PmGrds.

Constructors

GrdVec [PmGrd]

Guard tree language

newtype PmMatchGroup p

A guard tree denoting MatchGroup.

Constructors

PmMatchGroup (NonEmpty (PmMatch p))

data PmMatch p

A guard tree denoting Match: A payload describing the pats and a bunch of GRHS.

Constructors

PmMatch

Fields

pm_pats :: !p

pm_grhss :: !(PmGRHSs p)

data PmGRHSs p

A guard tree denoting GRHSs: A bunch of PmLet guards for local bindings from the GRHSss where clauses and the actual list of GRHS. See Note [Long-distance information for HsLocalBinds] in GHC.HsToCore.Pmc.Desugar.

Constructors

PmGRHSs

Fields

pgs_lcls :: !p

pgs_grhss :: !(NonEmpty (PmGRHS p))

data PmGRHS p

A guard tree denoting GRHS: A payload describing the grds and a SrcInfo useful for printing out in warnings messages.

Constructors

PmGRHS

Fields

pg_grds :: !p

pg_rhs :: !SrcInfo

newtype PmPatBind p

A guard tree denoting a pattern binding.

Constructors

PmPatBind (PmGRHS p)

newtype PmEmptyCase

A guard tree denoting an -XEmptyCase.

Constructors

PmEmptyCase

Fields

pe_var :: Id

Coverage Checking types

data RedSets

Redundancy sets, used to determine redundancy of RHSs and bang patterns (later digested into a CIRB).

Constructors

RedSets

Fields

rs_cov :: !Nablas

The Covered set; the set of values reaching a particular program point.

rs_div :: !Nablas

The Diverging set; empty if no match can lead to divergence. If it wasn't empty, we have to turn redundancy warnings into inaccessibility warnings for any subclauses.

rs_bangs :: !(OrdList (Nablas, SrcInfo))

If any of the Nablas is empty, the corresponding SrcInfo pin-points a bang pattern in source that is redundant. See Note [Dead bang patterns].

data Precision

Constructors

Approximate

Precise

data CheckResult a

Pattern-match coverage check result

Constructors

CheckResult

Fields

cr_ret :: !a

A hole for redundancy info and covered sets.

cr_uncov :: !Nablas

The set of uncovered values falling out at the bottom. (for -Wincomplete-patterns, but also important state for the algorithm)

cr_approx :: !Precision

A flag saying whether we ran into the maxPmCheckModels limit for the purpose of suggesting to crank it up in the warning message. Writer state.

Pre and post coverage checking synonyms

type Pre = GrdVec

Used as tree payload pre-checking. The LYG guards to check.

type Post = RedSets

Used as tree payload post-checking. The redundancy info we elaborated.

Normalised refinement types

module GHC.HsToCore.Pmc.Solver.Types