GHC.Iface.Binary

Description

Binary interface file support.

Public API for interface file serialisation

writeBinIface :: Profile -> TraceBinIFace -> FilePath -> ModIface -> IO ()

Write an interface file

readBinIface :: Profile -> NameCache -> CheckHiWay -> TraceBinIFace -> FilePath -> IO ModIface

Read an interface file.

readBinIfaceHeader :: Profile -> NameCache -> CheckHiWay -> TraceBinIFace -> FilePath -> IO (Fingerprint, BinHandle)

Read an interface file header, checking the magic number, version, and way. Returns the hash of the source file and a BinHandle which points at the start of the rest of the interface file data.

getSymtabName :: NameCache -> Dictionary -> SymbolTable -> BinHandle -> IO Name

data CheckHiWay

Constructors

CheckHiWay

IgnoreHiWay

data TraceBinIFace

Constructors

TraceBinIFace (SDoc -> IO ())

QuietBinIFace

getWithUserData :: Binary a => NameCache -> BinHandle -> IO a

This performs a get action after reading the dictionary and symbol table. It is necessary to run this before trying to deserialise any Names or FastStrings.

putWithUserData :: Binary a => TraceBinIFace -> BinHandle -> a -> IO ()

Put a piece of data with an initialised UserData field. This is necessary if you want to serialise Names or FastStrings. It also writes a symbol table and the dictionary. This segment should be read using getWithUserData.

Internal serialisation functions

getSymbolTable :: BinHandle -> NameCache -> IO SymbolTable

putName :: FSTable -> BinSymbolTable -> BinHandle -> Name -> IO ()

putSymbolTable :: BinHandle -> Int -> UniqFM Name (Int, Name) -> IO ()

data BinSymbolTable

Constructors

BinSymbolTable

Fields

bin_symtab_next :: !FastMutInt

bin_symtab_map :: !(IORef (UniqFM Name (Int, Name)))