GHC.Iface.Binary

Description

Binary interface file support.

Public API for interface file serialisation

writeBinIface :: Profile -> TraceBinIFace -> FilePath -> ModIface -> IO ()

Write an interface file

readBinIface :: CheckHiWay -> TraceBinIFace -> FilePath -> TcRnIf a b ModIface

Read an interface file

readBinIface_ :: Profile -> CheckHiWay -> TraceBinIFace -> FilePath -> NameCacheUpdater -> IO ModIface

Read an interface file in IO.

getSymtabName :: NameCacheUpdater -> Dictionary -> SymbolTable -> BinHandle -> IO Name

getDictFastString :: Dictionary -> BinHandle -> IO FastString

data CheckHiWay

Constructors

CheckHiWay

IgnoreHiWay

data TraceBinIFace

Constructors

TraceBinIFace (SDoc -> IO ())

QuietBinIFace

getWithUserData :: Binary a => NameCacheUpdater -> BinHandle -> IO a

This performs a get action after reading the dictionary and symbol table. It is necessary to run this before trying to deserialise any Names or FastStrings.

putWithUserData :: Binary a => TraceBinIFace -> BinHandle -> a -> IO ()

Put a piece of data with an initialised UserData field. This is necessary if you want to serialise Names or FastStrings. It also writes a symbol table and the dictionary. This segment should be read using getWithUserData.

Internal serialisation functions

getSymbolTable :: BinHandle -> NameCacheUpdater -> IO SymbolTable

putName :: BinDictionary -> BinSymbolTable -> BinHandle -> Name -> IO ()

putDictionary :: BinHandle -> Int -> UniqFM FastString (Int, FastString) -> IO ()

putFastString :: BinDictionary -> BinHandle -> FastString -> IO ()

putSymbolTable :: BinHandle -> Int -> UniqFM Name (Int, Name) -> IO ()

data BinSymbolTable

Constructors

BinSymbolTable

Fields

bin_symtab_next :: !FastMutInt

bin_symtab_map :: !(IORef (UniqFM Name (Int, Name)))

data BinDictionary

Constructors

BinDictionary

Fields

bin_dict_next :: !FastMutInt

bin_dict_map :: !(IORef (UniqFM FastString (Int, FastString)))