GHC.Iface.Recomp

Description

Module for detecting if recompilation is required

Documentation

checkOldIface :: HscEnv -> ModSummary -> SourceModified -> Maybe ModIface -> IO (RecompileRequired, Maybe ModIface)

Top level function to check if the version of an old interface file is equivalent to the current source file the user asked us to compile. If the same, we can avoid recompilation. We return a tuple where the first element is a bool saying if we should recompile the object file and the second is maybe the interface file, where Nothing means to rebuild the interface file and not use the existing one.

data RecompileRequired

Constructors

UpToDate
everything is up to date, recompilation is not required

MustCompile
The .hs file has been touched, or the .o/.hi file does not exist

RecompBecause String
The .o/.hi files are up to date, but something else has changed to force recompilation; the String says what (one-line summary)

recompileRequired :: RecompileRequired -> Bool

addFingerprints :: HscEnv -> PartialModIface -> IO ModIface

Add fingerprints for top-level declarations to a ModIface_.

See Note [Fingerprinting IfaceDecls]