GHC.JS.Ppr

Description

Pretty-printing JavaScript

Documentation

renderJs :: (JsToDoc a, JMacro a) => a -> Doc

Render a syntax tree as a pretty-printable document (simply showing the resultant doc produces a nice, well formatted String).

renderJs' :: (JsToDoc a, JMacro a) => RenderJs -> a -> Doc

renderPrefixJs :: (JsToDoc a, JMacro a) => FastString -> a -> Doc

Render a syntax tree as a pretty-printable document, using a given prefix to all generated names. Use this with distinct prefixes to ensure distinct generated names between independent calls to render(Prefix)Js.

renderPrefixJs' :: (JsToDoc a, JMacro a) => RenderJs -> FastString -> a -> Doc

class JsToDoc a where

Methods

jsToDocR :: RenderJs -> a -> Doc

defaultRenderJs :: RenderJs

data RenderJs

Constructors

RenderJs

Fields

renderJsS :: !(RenderJs -> JStat -> Doc)

renderJsE :: !(RenderJs -> JExpr -> Doc)

renderJsV :: !(RenderJs -> JVal -> Doc)

renderJsI :: !(RenderJs -> Ident -> Doc)

jsToDoc :: JsToDoc a => a -> Doc

pprStringLit :: FastString -> Doc

flattenBlocks :: [JStat] -> [JStat]

braceNest :: Doc -> Doc

hangBrace :: Doc -> Doc -> Doc

Hang with braces:

hdr { body }

Orphan instances

Outputable JExpr

Instance details

Methods

ppr :: JExpr -> SDoc

Outputable JVal

Instance details

Methods

ppr :: JVal -> SDoc