GHC.Linker.Unit

Description

Linking Haskell units

Documentation

collectLinkOpts :: DynFlags -> [UnitInfo] -> ([String], [String], [String])

collectArchives :: DynFlags -> UnitInfo -> IO [FilePath]

getUnitLinkOpts :: DynFlags -> UnitEnv -> [UnitId] -> IO ([String], [String], [String])

Find all the link options in these and the preload packages, returning (package hs lib options, extra library options, other flags)

getLibs :: DynFlags -> UnitEnv -> [UnitId] -> IO [(String, String)]