GHC.Parser.Annotation

Core Exact Print Annotation types

data EpToken (tok :: Symbol)

A token stored in the syntax tree. For example, when parsing a let-expression, we store EpToken "let" and EpToken "in". The locations of those tokens can be used to faithfully reproduce (exactprint) the original program text.

Constructors

NoEpTok

EpTok !EpaLocation

data EpUniToken (tok :: Symbol) (utok :: Symbol)

With UnicodeSyntax, there might be multiple ways to write the same token. For example an arrow could be either -> or →. This choice must be recorded in order to exactprint such tokens, so instead of EpToken "->" we introduce EpUniToken "->" "→".

Constructors

NoEpUniTok

EpUniTok !EpaLocation !IsUnicodeSyntax

getEpTokenSrcSpan :: forall (tok :: Symbol). EpToken tok -> SrcSpan

getEpTokenLocs :: forall (tok :: Symbol). [EpToken tok] -> [EpaLocation]

getEpTokenLoc :: forall (tok :: Symbol). EpToken tok -> EpaLocation

getEpUniTokenLoc :: forall (tok :: Symbol) (toku :: Symbol). EpUniToken tok toku -> EpaLocation

type TokDcolon = EpUniToken "::" "\8759"

type TokDarrow = EpUniToken "=>" "\8658"

type TokRarrow = EpUniToken "->" "\8594"

type TokForall = EpUniToken "forall" "\8704"

data EpLayout

Layout information for declarations.

Constructors

EpExplicitBraces !(EpToken "{") !(EpToken "}")
Explicit braces written by the user.

class C a where { foo :: a; bar :: a }

EpVirtualBraces
Virtual braces inserted by the layout algorithm.

class C a where
  foo :: a
  bar :: a

Fields

!Int

Layout column (indentation level, begins at 1)

EpNoLayout
Empty or compiler-generated blocks do not have layout information associated with them.

data EpaComment

Constructors

EpaComment

Fields

ac_tok :: EpaCommentTok

ac_prior_tok :: RealSrcSpan

The location of the prior token, used in exact printing. The EpaComment appears as an LEpaComment containing its location. The difference between the end of the prior token and the start of this location is used for the spacing when exact printing the comment.

data EpaCommentTok

Constructors

EpaDocComment HsDocString
a docstring that can be pretty printed using pprHsDocString

EpaDocOptions String
doc options (prune, ignore-exports, etc)

EpaLineComment String
comment starting by "--"

EpaBlockComment String
comment in {- -}

data IsUnicodeSyntax

Certain tokens can have alternate representations when unicode syntax is enabled. This flag is attached to those tokens in the lexer so that the original source representation can be reproduced in the corresponding EpAnnotation

Constructors

UnicodeSyntax

NormalSyntax

data HasE

Some template haskell tokens have two variants, one with an e the other not:

 [| or [e|
 [|| or [e||

This type indicates whether the e is present or not.

Constructors

HasE

NoE

In-tree Exact Print Annotations

type EpaLocation = EpaLocation' [LEpaComment]

data EpaLocation' a

The anchor for an exact print annotation. The Parser inserts the EpaSpan variant, giving the exact location of the original item in the parsed source. This can be replaced by the EpaDelta version, to provide a position for the item relative to the end of the previous item in the source. This is useful when editing an AST prior to exact printing the changed one. The EpaDelta also contains the original SrcSpan for use by tools wanting to manipulate the AST after converting it using ghc-exactprint' makeDeltaAst.

Constructors

EpaSpan !SrcSpan

EpaDelta !SrcSpan !DeltaPos !a

epaLocationRealSrcSpan :: EpaLocation' a -> RealSrcSpan

Used in the parser only, extract the RealSrcSpan from an EpaLocation. The parser will never insert a DeltaPos, so the partial function is safe.

data TokenLocation

Tokens embedded in the AST have an EpaLocation, unless they come from generated code (e.g. by TH).

Constructors

NoTokenLoc

TokenLoc !EpaLocation

data DeltaPos

Spacing between output items when exact printing. It captures the spacing from the current print position on the page to the position required for the thing about to be printed. This is either on the same line in which case is is simply the number of spaces to emit, or it is some number of lines down, with a given column offset. The exact printing algorithm keeps track of the column offset pertaining to the current anchor position, so the deltaColumn is the additional spaces to add in this case. See https://gitlab.haskell.org/ghc/ghc/wikis/api-annotations for details.

Constructors

SameLine

Fields

deltaColumn :: !Int

DifferentLine

Fields

deltaLine :: !Int

deltaLine should always be > 0

deltaColumn :: !Int

deltaPos :: Int -> Int -> DeltaPos

Smart constructor for a DeltaPos. It preserves the invariant that for the DifferentLine constructor deltaLine is always > 0.

getDeltaLine :: DeltaPos -> Int

data EpAnn ann

The exact print annotations (EPAs) are kept in the HsSyn AST for the GhcPs phase. They are usually inserted into the AST by the parser, and in case of generated code (e.g. by TemplateHaskell) they are usually initialized using NoAnn type class.

A goal of the annotations is that an AST can be edited, including moving subtrees from one place to another, duplicating them, and so on. This means that each fragment must be self-contained. To this end, each annotated fragment keeps track of the anchor position it was originally captured at, being simply the start span of the topmost element of the ast fragment. This gives us a way to later re-calculate all Located items in this layer of the AST, as well as any annotations captured. The comments associated with the AST fragment are also captured here.

The ann type parameter allows this general structure to be specialised to the specific set of locations of original exact print annotation elements. For example

type SrcSpannAnnA = EpAnn AnnListItem

is a commonly used type alias that specializes the ann type parameter to AnnListItem.

The spacing between the items under the scope of a given EpAnn is normally derived from the original Anchor. But if a sub-element is not in its original position, the required spacing can be captured using an appropriate EpaDelta value for the entry Anchor. This allows us to freely move elements around, and stitch together new AST fragments out of old ones, and have them still printed out in a precise way.

Constructors

EpAnn

Fields

entry :: !EpaLocation

Base location for the start of the syntactic element holding the annotations.

anns :: !ann

Annotations added by the Parser

comments :: !EpAnnComments

Comments enclosed in the SrcSpan of the element this EpAnn is attached to

spanAsAnchor :: SrcSpan -> EpaLocation' a

realSpanAsAnchor :: RealSrcSpan -> EpaLocation' a

noSpanAnchor :: NoAnn a => EpaLocation' a

class NoAnn a where

Methods

noAnn :: a

equivalent of mempty, but does not need Semigroup

Comments in Annotations

data EpAnnComments

When we are parsing we add comments that belong to a particular AST element, and print them together with the element, interleaving them into the output stream. But when editing the AST to move fragments around it is useful to be able to first separate the comments into those occurring before the AST element and those following it. The EpaCommentsBalanced constructor is used to do this. The GHC parser will only insert the EpaComments form.

Constructors

EpaComments

Fields

priorComments :: ![LEpaComment]

EpaCommentsBalanced

Fields

priorComments :: ![LEpaComment]

followingComments :: ![LEpaComment]

type LEpaComment = GenLocated NoCommentsLocation EpaComment

type NoCommentsLocation = EpaLocation' NoComments

data NoComments

Constructors

NoComments

emptyComments :: EpAnnComments

epaToNoCommentsLocation :: EpaLocation -> NoCommentsLocation

noCommentsToEpaLocation :: NoCommentsLocation -> EpaLocation

getFollowingComments :: EpAnnComments -> [LEpaComment]

setFollowingComments :: EpAnnComments -> [LEpaComment] -> EpAnnComments

setPriorComments :: EpAnnComments -> [LEpaComment] -> EpAnnComments

type EpAnnCO

Arguments

= EpAnn NoEpAnns
Api Annotations for comments only

Annotations in GenLocated

type LocatedA = GenLocated SrcSpanAnnA

type LocatedL = GenLocated SrcSpanAnnL

type LocatedC = GenLocated SrcSpanAnnC

type LocatedN = GenLocated SrcSpanAnnN

type LocatedAn an = GenLocated (EpAnn an)

General representation of a GenLocated type carrying a parameterised annotation type.

type LocatedP = GenLocated SrcSpanAnnP

type LocatedLC = GenLocated SrcSpanAnnLC

type LocatedLS = GenLocated SrcSpanAnnLS

type LocatedLW = GenLocated SrcSpanAnnLW

type LocatedLI = GenLocated SrcSpanAnnLI

type SrcSpanAnnA = EpAnn AnnListItem

type SrcSpanAnnL = EpAnn (AnnList ())

type SrcSpanAnnP = EpAnn AnnPragma

type SrcSpanAnnC = EpAnn AnnContext

type SrcSpanAnnN = EpAnn NameAnn

type SrcSpanAnnLC = EpAnn (AnnList [EpToken ","])

type SrcSpanAnnLW = EpAnn (AnnList (EpToken "where"))

type SrcSpanAnnLS = EpAnn (AnnList ())

type SrcSpanAnnLI = EpAnn (AnnList (EpToken "hiding", [EpToken ","]))

type LocatedE = GenLocated EpaLocation

Annotation data types used in GenLocated

data AnnListItem

Annotation for items appearing in a list. They can have one or more trailing punctuations items, such as commas or semicolons.

Constructors

AnnListItem

Fields

lann_trailing :: [TrailingAnn]

data AnnList a

Annotation for the "container" of a list. This captures surrounding items such as braces if present, and introductory keywords such as 'where'.

Constructors

AnnList

Fields

al_anchor :: !(Maybe EpaLocation)

start point of a list having layout

al_brackets :: !AnnListBrackets

al_semis :: [EpToken ";"]

al_rest :: !a

al_trailing :: ![TrailingAnn]

items appearing after the list, such as '=>' for a context

data AnnListBrackets

Constructors

ListParens (EpToken "(") (EpToken ")")

ListBraces (EpToken "{") (EpToken "}")

ListSquare (EpToken "[") (EpToken "]")

ListBanana (EpUniToken "(|" "\10631") (EpUniToken "|)" "\10632")

ListNone

data AnnParen

exact print annotation for an item having surrounding "brackets", such as tuples or lists

Constructors

AnnParens (EpToken "(") (EpToken ")")
'(', ')'

AnnParensHash (EpToken "(#") (EpToken "#)")
'(#', '#)'

AnnParensSquare (EpToken "[") (EpToken "]")
'[', ']'

data AnnPragma

exact print annotation used for capturing the locations of annotations in pragmas.

Constructors

AnnPragma

Fields

apr_open :: EpaLocation

apr_close :: EpToken "#-}"

apr_squares :: (EpToken "[", EpToken "]")

apr_loc1 :: EpaLocation

apr_loc2 :: EpaLocation

apr_type :: EpToken "type"

apr_module :: EpToken "module"

data AnnContext

Exact print annotation for the Context data type.

Constructors

AnnContext

Fields

ac_darrow :: Maybe TokDarrow

location of the '=>', if present.

ac_open :: [EpToken "("]

zero or more opening parentheses.

ac_close :: [EpToken ")"]

zero or more closing parentheses.

data NameAnn

exact print annotations for a RdrName. There are many kinds of adornment that can be attached to a given RdrName. This type captures them, as detailed on the individual constructors.

Constructors

NameAnn
Used for a name with an adornment, so `foo`, (bar)

Fields

nann_adornment :: NameAdornment

nann_name :: EpaLocation

nann_trailing :: [TrailingAnn]

NameAnnCommas
Used for (,,,), or ()

Fields

nann_adornment :: NameAdornment

nann_commas :: [EpToken ","]

nann_trailing :: [TrailingAnn]

NameAnnBars
Used for (# | | #)

Fields

nann_parensh :: (EpToken "(#", EpToken "#)")

nann_bars :: [EpToken "|"]

nann_trailing :: [TrailingAnn]

NameAnnOnly
Used for (), (##), []

Fields

nann_adornment :: NameAdornment

nann_trailing :: [TrailingAnn]

NameAnnRArrow
Used for ->, as an identifier

Fields

nann_mopen :: Maybe (EpToken "(")

nann_arrow :: TokRarrow

nann_mclose :: Maybe (EpToken ")")

nann_trailing :: [TrailingAnn]

NameAnnQuote
Used for an item with a leading '. The annotation for unquoted item is stored in nann_quoted.

Fields

nann_quote :: EpToken "'"

nann_quoted :: SrcSpanAnnN

nann_trailing :: [TrailingAnn]

NameAnnTrailing
Used when adding a TrailingAnn to an existing LocatedN which has no Api Annotation.

Fields

nann_trailing :: [TrailingAnn]

data NameAdornment

A NameAnn can capture the locations of surrounding adornments, such as parens or backquotes. This data type identifies what particular pair are being used.

Constructors

NameParens (EpToken "(") (EpToken ")")

NameParensHash (EpToken "(#") (EpToken "#)")

NameBackquotes (EpToken "`") (EpToken "`")

NameSquare (EpToken "[") (EpToken "]")

NameNoAdornment

data NoEpAnns

Constructors

NoEpAnns

data AnnSortKey tag

Captures the sort order of sub elements for ValBinds, ClassDecl, ClsInstDecl

Constructors

NoAnnSortKey

AnnSortKey [tag]

data DeclTag

Used to track interleaving of class methods, class signatures, associated types and associate type defaults in ClassDecl and ClsInstDecl.

Constructors

ClsMethodTag

ClsSigTag

ClsAtTag

ClsAtdTag

data BindTag

Used to track of interleaving of binds and signatures for ValBind

Constructors

BindTag

SigDTag

Trailing annotations in lists

data TrailingAnn

Captures the location of punctuation occurring between items, normally in a list. It is captured as a trailing annotation.

Constructors

AddSemiAnn (EpToken ";")
Trailing ';'

AddCommaAnn (EpToken ",")
Trailing ','

AddVbarAnn (EpToken "|")
Trailing '|'

AddDarrowAnn TokDarrow
Trailing '=>' / ⇒

ta_location :: TrailingAnn -> EpaLocation

addTrailingAnnToA :: TrailingAnn -> EpAnnComments -> EpAnn AnnListItem -> EpAnn AnnListItem

Helper function used in the parser to add a TrailingAnn items to an existing annotation.

addTrailingAnnToL :: TrailingAnn -> EpAnnComments -> EpAnn (AnnList a) -> EpAnn (AnnList a)

Helper function used in the parser to add a TrailingAnn items to an existing annotation.

addTrailingCommaToN :: EpAnn NameAnn -> EpaLocation -> EpAnn NameAnn

Helper function used in the parser to add a comma location to an existing annotation.

noTrailingN :: SrcSpanAnnN -> SrcSpanAnnN

Utilities for converting between different GenLocated when

we do not care about the annotations.

l2l :: (HasLoc a, HasAnnotation b) => a -> b

Helper function for converting annotation types. Discards any annotations

la2la :: (HasLoc l, HasAnnotation l2) => GenLocated l a -> GenLocated l2 a

Helper function for converting annotation types. Discards any annotations

reLoc :: (HasLoc (GenLocated a e), HasAnnotation b) => GenLocated a e -> GenLocated b e

class HasLoc a where

Methods

getHasLoc :: a -> SrcSpan

conveniently calculate locations for things without locations attached

getHasLocList :: HasLoc a => [a] -> SrcSpan

srcSpan2e :: SrcSpan -> EpaLocation

realSrcSpan :: SrcSpan -> RealSrcSpan

Building up annotations

reAnnL :: ann -> EpAnnComments -> Located e -> GenLocated (EpAnn ann) e

reAnnC :: AnnContext -> EpAnnComments -> Located a -> LocatedC a

addAnnsA :: SrcSpanAnnA -> [TrailingAnn] -> EpAnnComments -> SrcSpanAnnA

widenSpanL :: SrcSpan -> [EpaLocation] -> SrcSpan

The annotations need to all come after the anchor. Make sure this is the case.

widenSpanT :: forall (tok :: Symbol). SrcSpan -> EpToken tok -> SrcSpan

widenAnchorT :: forall (tok :: Symbol). EpaLocation -> EpToken tok -> EpaLocation

widenAnchorS :: EpaLocation -> SrcSpan -> EpaLocation

widenLocatedAnL :: EpAnn an -> [EpaLocation] -> EpAnn an

listLocation :: [LocatedAn an a] -> EpaLocation

Querying annotations

getLocAnn :: Located a -> SrcSpanAnnA

epAnnComments :: EpAnn an -> EpAnnComments

Working with locations of annotations

sortLocatedA :: HasLoc (EpAnn a) => [GenLocated (EpAnn a) e] -> [GenLocated (EpAnn a) e]

mapLocA :: NoAnn ann => (a -> b) -> GenLocated SrcSpan a -> GenLocated (EpAnn ann) b

combineLocsA :: Semigroup a => GenLocated (EpAnn a) e1 -> GenLocated (EpAnn a) e2 -> EpAnn a

combineSrcSpansA :: Semigroup a => EpAnn a -> EpAnn a -> EpAnn a

addCLocA :: (HasLoc a, HasLoc b, HasAnnotation l) => a -> b -> c -> GenLocated l c

Combine locations from two Located things and add them to a third thing

Constructing GenLocated annotation types when we do not care

class HasAnnotation e where

Methods

noAnnSrcSpan :: SrcSpan -> e

locA :: HasLoc a => a -> SrcSpan

noLocA :: HasAnnotation e => a -> GenLocated e a

getLocA :: HasLoc a => GenLocated a e -> SrcSpan

noSrcSpanA :: HasAnnotation e => e

Working with comments in annotations

noComments :: EpAnnCO

comment :: RealSrcSpan -> EpAnnComments -> EpAnnCO

addCommentsToEpAnn :: NoAnn ann => EpAnn ann -> EpAnnComments -> EpAnn ann

Add additional comments to a EpAnn, used for manipulating the AST prior to exact printing the changed one.

setCommentsEpAnn :: NoAnn ann => EpAnn ann -> EpAnnComments -> EpAnn ann

Replace any existing comments on a EpAnn, used for manipulating the AST prior to exact printing the changed one.

transferAnnsA :: SrcSpanAnnA -> SrcSpanAnnA -> (SrcSpanAnnA, SrcSpanAnnA)

Transfer comments and trailing items from the annotations in the first SrcSpanAnnA argument to those in the second.

transferAnnsOnlyA :: SrcSpanAnnA -> SrcSpanAnnA -> (SrcSpanAnnA, SrcSpanAnnA)

Transfer trailing items from the annotations in the first SrcSpanAnnA argument to those in the second.

transferCommentsOnlyA :: EpAnn a -> EpAnn b -> (EpAnn a, EpAnn b)

Transfer comments from the annotations in the first SrcSpanAnnA argument to those in the second.

transferPriorCommentsA :: SrcSpanAnnA -> SrcSpanAnnA -> (SrcSpanAnnA, SrcSpanAnnA)

Transfer prior comments only from the annotations in the first SrcSpanAnnA argument to those in the second.

transferFollowingA :: SrcSpanAnnA -> SrcSpanAnnA -> (SrcSpanAnnA, SrcSpanAnnA)

Transfer trailing items but not comments from the annotations in the first SrcSpanAnnA argument to those in the second.

commentsOnlyA :: NoAnn ann => EpAnn ann -> EpAnn ann

Remove the exact print annotations payload, leaving only the anchor and comments.

removeCommentsA :: EpAnn ann -> EpAnn ann

Remove the comments, leaving the exact print annotations payload

placeholderRealSpan :: RealSrcSpan