GHC.Parser.Annotation

Core Exact Print Annotation types

data AnnKeywordId

Exact print annotations exist so that tools can perform source to source conversions of Haskell code. They are used to keep track of the various syntactic keywords that are not otherwise captured in the AST.

The wiki page describing this feature is https://gitlab.haskell.org/ghc/ghc/wikis/api-annotations https://gitlab.haskell.org/ghc/ghc/-/wikis/implementing-trees-that-grow/in-tree-api-annotations

Note: in general the names of these are taken from the corresponding token, unless otherwise noted See note [exact print annotations] above for details of the usage

Constructors

AnnAnyclass

AnnAs

AnnAt

AnnBang
!

AnnBackquote
'`'

AnnBy

AnnCase
case or lambda case

AnnClass

AnnClose
'#)' or '#-}' etc

AnnCloseB
'|)'

AnnCloseBU
'|)', unicode variant

AnnCloseC
'}'

AnnCloseQ
'|]'

AnnCloseQU
'|]', unicode variant

AnnCloseP
')'

AnnClosePH
'#)'

AnnCloseS
']'

AnnColon

AnnComma
as a list separator

AnnCommaTuple
in a RdrName for a tuple

AnnDarrow
'=>'

AnnDarrowU
'=>', unicode variant

AnnData

AnnDcolon
'::'

AnnDcolonU
'::', unicode variant

AnnDefault

AnnDeriving

AnnDo

AnnDot
.

AnnDotdot
'..'

AnnElse

AnnEqual

AnnExport

AnnFamily

AnnForall

AnnForallU
Unicode variant

AnnForeign

AnnFunId
for function name in matches where there are multiple equations for the function.

AnnGroup

AnnHeader
for CType

AnnHiding

AnnIf

AnnImport

AnnIn

AnnInfix
'infix' or 'infixl' or 'infixr'

AnnInstance

AnnLam

AnnLarrow
'<-'

AnnLarrowU
'<-', unicode variant

AnnLet

AnnLollyU
The ⊸ unicode arrow

AnnMdo

AnnMinus
-

AnnModule

AnnNewtype

AnnName
where a name loses its location in the AST, this carries it

AnnOf

AnnOpen
'{-# DEPRECATED' etc. Opening of pragmas where the capitalisation of the string can be changed by the user. The actual text used is stored in a SourceText on the relevant pragma item.

AnnOpenB
'(|'

AnnOpenBU
'(|', unicode variant

AnnOpenC
'{'

AnnOpenE
'[e|' or '[e||'

AnnOpenEQ
'[|'

AnnOpenEQU
'[|', unicode variant

AnnOpenP
'('

AnnOpenS
'['

AnnOpenPH
'(#'

AnnDollar
prefix $ -- TemplateHaskell

AnnDollarDollar
prefix $$ -- TemplateHaskell

AnnPackageName

AnnPattern

AnnPercent
% -- for HsExplicitMult

AnnPercentOne
'%1' -- for HsLinearArrow

AnnProc

AnnQualified

AnnRarrow
->

AnnRarrowU
->, unicode variant

AnnRec

AnnRole

AnnSafe

AnnSemi
';'

AnnSimpleQuote
'''

AnnSignature

AnnStatic
static

AnnStock

AnnThen

AnnThTyQuote
double '''

AnnTilde
~

AnnType

AnnUnit
() for types

AnnUsing

AnnVal
e.g. INTEGER

AnnValStr
String value, will need quotes when output

AnnVbar
'|'

AnnVia
via

AnnWhere

Annlarrowtail
-<

AnnlarrowtailU
-<, unicode variant

Annrarrowtail
->

AnnrarrowtailU
->, unicode variant

AnnLarrowtail
-<<

AnnLarrowtailU
-<<, unicode variant

AnnRarrowtail
>>-

AnnRarrowtailU
>>-, unicode variant

data EpaComment

Constructors

EpaComment

Fields

ac_tok :: EpaCommentTok

ac_prior_tok :: RealSrcSpan

The location of the prior token, used in exact printing. The EpaComment appears as an LEpaComment containing its location. The difference between the end of the prior token and the start of this location is used for the spacing when exact printing the comment.

data EpaCommentTok

Constructors

EpaDocCommentNext String
something beginning '-- |'

EpaDocCommentPrev String
something beginning '-- ^'

EpaDocCommentNamed String
something beginning '-- $'

EpaDocSection Int String
a section heading

EpaDocOptions String
doc options (prune, ignore-exports, etc)

EpaLineComment String
comment starting by "--"

EpaBlockComment String
comment in {- -}

EpaEofComment
empty comment, capturing location of EOF

data IsUnicodeSyntax

Certain tokens can have alternate representations when unicode syntax is enabled. This flag is attached to those tokens in the lexer so that the original source representation can be reproduced in the corresponding EpAnnotation

Constructors

UnicodeSyntax

NormalSyntax

unicodeAnn :: AnnKeywordId -> AnnKeywordId

Convert a normal annotation into its unicode equivalent one

data HasE

Some template haskell tokens have two variants, one with an e the other not:

 [| or [e|
 [|| or [e||

This type indicates whether the e is present or not.

Constructors

HasE

NoE

In-tree Exact Print Annotations

data AddEpAnn

Captures an annotation, storing the AnnKeywordId and its location. The parser only ever inserts EpaLocation fields with a RealSrcSpan being the original location of the annotation in the source file. The EpaLocation can also store a delta position if the AST has been modified and needs to be pretty printed again. The usual way an AddEpAnn is created is using the mj ("make jump") function, and then it can be inserted into the appropriate annotation.

Constructors

AddEpAnn AnnKeywordId EpaLocation

data EpaLocation

The anchor for an AnnKeywordId. The Parser inserts the EpaSpan variant, giving the exact location of the original item in the parsed source. This can be replaced by the EpaDelta version, to provide a position for the item relative to the end of the previous item in the source. This is useful when editing an AST prior to exact printing the changed one. The list of comments in the EpaDelta variant captures any comments between the prior output and the thing being marked here, since we cannot otherwise sort the relative order.

Constructors

EpaSpan !RealSrcSpan

EpaDelta !DeltaPos ![LEpaComment]

epaLocationRealSrcSpan :: EpaLocation -> RealSrcSpan

Used in the parser only, extract the SrcSpan from an EpaLocation. The parser will never insert a DeltaPos, so the partial function is safe.

epaLocationFromSrcAnn :: SrcAnn ann -> EpaLocation

data DeltaPos

Spacing between output items when exact printing. It captures the spacing from the current print position on the page to the position required for the thing about to be printed. This is either on the same line in which case is is simply the number of spaces to emit, or it is some number of lines down, with a given column offset. The exact printing algorithm keeps track of the column offset pertaining to the current anchor position, so the deltaColumn is the additional spaces to add in this case. See https://gitlab.haskell.org/ghc/ghc/wikis/api-annotations for details.

Constructors

SameLine

Fields

deltaColumn :: !Int

DifferentLine

Fields

deltaLine :: !Int

deltaLine should always be > 0

deltaColumn :: !Int

deltaPos :: Int -> Int -> DeltaPos

Smart constructor for a DeltaPos. It preserves the invariant that for the DifferentLine constructor deltaLine is always > 0.

getDeltaLine :: DeltaPos -> Int

data EpAnn ann

The exact print annotations (EPAs) are kept in the HsSyn AST for the GhcPs phase. We do not always have EPAs though, only for code that has been parsed as they do not exist for generated code. This type captures that they may be missing.

A goal of the annotations is that an AST can be edited, including moving subtrees from one place to another, duplicating them, and so on. This means that each fragment must be self-contained. To this end, each annotated fragment keeps track of the anchor position it was originally captured at, being simply the start span of the topmost element of the ast fragment. This gives us a way to later re-calculate all Located items in this layer of the AST, as well as any annotations captured. The comments associated with the AST fragment are also captured here.

The ann type parameter allows this general structure to be specialised to the specific set of locations of original exact print annotation elements. So for HsLet we have

type instance XLet GhcPs = EpAnn AnnsLet data AnnsLet = AnnsLet { alLet :: EpaLocation, alIn :: EpaLocation } deriving Data

The spacing between the items under the scope of a given EpAnn is normally derived from the original Anchor. But if a sub-element is not in its original position, the required spacing can be directly captured in the anchor_op field of the entry Anchor. This allows us to freely move elements around, and stitch together new AST fragments out of old ones, and have them still printed out in a precise way.

Constructors

EpAnn

Fields

entry :: !Anchor

Base location for the start of the syntactic element holding the annotations.

anns :: !ann

Annotations added by the Parser

comments :: !EpAnnComments

Comments enclosed in the SrcSpan of the element this EpAnn is attached to

EpAnnNotUsed
No Annotation for generated code, e.g. from TH, deriving, etc.

data Anchor

An Anchor records the base location for the start of the syntactic element holding the annotations, and is used as the point of reference for calculating delta positions for contained annotations. It is also normally used as the reference point for the spacing of the element relative to its container. If it is moved, that relationship is tracked in the anchor_op instead.

Constructors

Anchor

Fields

anchor :: RealSrcSpan

Base location for the start of the syntactic element holding the annotations.

anchor_op :: AnchorOperation

data AnchorOperation

If tools modify the parsed source, the MovedAnchor variant can directly provide the spacing for this item relative to the previous one when printing. This allows AST fragments with a particular anchor to be freely moved, without worrying about recalculating the appropriate anchor span.

Constructors

UnchangedAnchor

MovedAnchor DeltaPos

spanAsAnchor :: SrcSpan -> Anchor

realSpanAsAnchor :: RealSrcSpan -> Anchor

noAnn :: EpAnn a

Short form for EpAnnNotUsed

Comments in Annotations

data EpAnnComments

When we are parsing we add comments that belong a particular AST element, and print them together with the element, interleaving them into the output stream. But when editing the AST to move fragments around it is useful to be able to first separate the comments into those occuring before the AST element and those following it. The EpaCommentsBalanced constructor is used to do this. The GHC parser will only insert the EpaComments form.

Constructors

EpaComments

Fields

priorComments :: ![LEpaComment]

EpaCommentsBalanced

Fields

priorComments :: ![LEpaComment]

followingComments :: ![LEpaComment]

type LEpaComment = GenLocated Anchor EpaComment

emptyComments :: EpAnnComments

getFollowingComments :: EpAnnComments -> [LEpaComment]

setFollowingComments :: EpAnnComments -> [LEpaComment] -> EpAnnComments

setPriorComments :: EpAnnComments -> [LEpaComment] -> EpAnnComments

type EpAnnCO

Arguments

= EpAnn NoEpAnns
Api Annotations for comments only

Annotations in GenLocated

type LocatedA = GenLocated SrcSpanAnnA

type LocatedL = GenLocated SrcSpanAnnL

type LocatedC = GenLocated SrcSpanAnnC

type LocatedN = GenLocated SrcSpanAnnN

type LocatedAn an = GenLocated (SrcAnn an)

General representation of a GenLocated type carrying a parameterised annotation type.

type LocatedP = GenLocated SrcSpanAnnP

type SrcSpanAnnA = SrcAnn AnnListItem

type SrcSpanAnnL = SrcAnn AnnList

type SrcSpanAnnP = SrcAnn AnnPragma

type SrcSpanAnnC = SrcAnn AnnContext

type SrcSpanAnnN = SrcAnn NameAnn

data SrcSpanAnn' a

The 'SrcSpanAnn'' type wraps a normal SrcSpan, together with an extra annotation type. This is mapped to a specific GenLocated usage in the AST through the XRec and Anno type families.

Constructors

SrcSpanAnn

Fields

ann :: !a

locA :: !SrcSpan

type SrcAnn ann = SrcSpanAnn' (EpAnn ann)

We mostly use 'SrcSpanAnn'' with an 'EpAnn''

Annotation data types used in GenLocated

data AnnListItem

Annotation for items appearing in a list. They can have one or more trailing punctuations items, such as commas or semicolons.

Constructors

AnnListItem

Fields

lann_trailing :: [TrailingAnn]

data AnnList

Annotation for the "container" of a list. This captures surrounding items such as braces if present, and introductory keywords such as 'where'.

Constructors

AnnList

Fields

al_anchor :: Maybe Anchor

start point of a list having layout

al_open :: Maybe AddEpAnn

al_close :: Maybe AddEpAnn

al_rest :: [AddEpAnn]

context, such as 'where' keyword

al_trailing :: [TrailingAnn]

items appearing after the list, such as '=>' for a context

data AnnParen

exact print annotation for an item having surrounding "brackets", such as tuples or lists

Constructors

AnnParen

Fields

ap_adornment :: ParenType

ap_open :: EpaLocation

ap_close :: EpaLocation

data ParenType

Detail of the "brackets" used in an AnnParen exact print annotation.

Constructors

AnnParens
'(', ')'

AnnParensHash
'(#', '#)'

AnnParensSquare
'[', ']'

parenTypeKws :: ParenType -> (AnnKeywordId, AnnKeywordId)

Maps the ParenType to the related opening and closing AnnKeywordId. Used when actually printing the item.

data AnnPragma

exact print annotation used for capturing the locations of annotations in pragmas.

Constructors

AnnPragma

Fields

apr_open :: AddEpAnn

apr_close :: AddEpAnn

apr_rest :: [AddEpAnn]

data AnnContext

Exact print annotation for the Context data type.

Constructors

AnnContext

Fields

ac_darrow :: Maybe (IsUnicodeSyntax, EpaLocation)

location and encoding of the '=>', if present.

ac_open :: [EpaLocation]

zero or more opening parentheses.

ac_close :: [EpaLocation]

zero or more closing parentheses.

data NameAnn

exact print annotations for a RdrName. There are many kinds of adornment that can be attached to a given RdrName. This type captures them, as detailed on the individual constructors.

Constructors

NameAnn
Used for a name with an adornment, so `foo`, (bar)

Fields

nann_adornment :: NameAdornment

nann_open :: EpaLocation

nann_name :: EpaLocation

nann_close :: EpaLocation

nann_trailing :: [TrailingAnn]

NameAnnCommas
Used for (,,,), or @()#

Fields

nann_adornment :: NameAdornment

nann_open :: EpaLocation

nann_commas :: [EpaLocation]

nann_close :: EpaLocation

nann_trailing :: [TrailingAnn]

NameAnnOnly
Used for (), (##), []

Fields

nann_adornment :: NameAdornment

nann_open :: EpaLocation

nann_close :: EpaLocation

nann_trailing :: [TrailingAnn]

NameAnnRArrow
Used for ->, as an identifier

Fields

nann_name :: EpaLocation

nann_trailing :: [TrailingAnn]

NameAnnQuote
Used for an item with a leading '. The annotation for unquoted item is stored in nann_quoted.

Fields

nann_quote :: EpaLocation

nann_quoted :: SrcSpanAnnN

nann_trailing :: [TrailingAnn]

NameAnnTrailing
Used when adding a TrailingAnn to an existing LocatedN which has no Api Annotation (via the EpAnnNotUsed constructor.

Fields

nann_trailing :: [TrailingAnn]

data NameAdornment

A NameAnn can capture the locations of surrounding adornments, such as parens or backquotes. This data type identifies what particular pair are being used.

Constructors

NameParens
'(' ')'

NameParensHash
'(#' '#)'

NameBackquotes
'`'

NameSquare
'[' ']'

data NoEpAnns

Constructors

NoEpAnns

data AnnSortKey

Captures the sort order of sub elements. This is needed when the sub-elements have been split (as in a HsLocalBind which holds separate binds and sigs) or for infix patterns where the order has been re-arranged. It is captured explicitly so that after the Delta phase a SrcSpan is used purely as an index into the annotations, allowing transformations of the AST including the introduction of new Located items or re-arranging existing ones.

Constructors

NoAnnSortKey

AnnSortKey [RealSrcSpan]

Trailing annotations in lists

data TrailingAnn

Captures the location of punctuation occuring between items, normally in a list. It is captured as a trailing annotation.

Constructors

AddSemiAnn EpaLocation
Trailing ';'

AddCommaAnn EpaLocation
Trailing ','

AddVbarAnn EpaLocation
Trailing '|'

AddRarrowAnn EpaLocation
Trailing ->

AddRarrowAnnU EpaLocation
Trailing ->, unicode variant

AddLollyAnnU EpaLocation
Trailing ⊸

addTrailingAnnToA :: SrcSpan -> TrailingAnn -> EpAnnComments -> EpAnn AnnListItem -> EpAnn AnnListItem

Helper function used in the parser to add a TrailingAnn items to an existing annotation.

addTrailingAnnToL :: SrcSpan -> TrailingAnn -> EpAnnComments -> EpAnn AnnList -> EpAnn AnnList

Helper function used in the parser to add a TrailingAnn items to an existing annotation.

addTrailingCommaToN :: SrcSpan -> EpAnn NameAnn -> EpaLocation -> EpAnn NameAnn

Helper function used in the parser to add a comma location to an existing annotation.

Utilities for converting between different GenLocated when

we do not care about the annotations.

la2na :: SrcSpanAnn' a -> SrcSpanAnnN

Helper function (temporary) during transition of names Discards any annotations

na2la :: SrcSpanAnn' a -> SrcAnn ann

Helper function (temporary) during transition of names Discards any annotations

n2l :: LocatedN a -> LocatedA a

l2n :: LocatedAn a1 a2 -> LocatedN a2

Helper function (temporary) during transition of names Discards any annotations

l2l :: SrcSpanAnn' a -> SrcAnn ann

la2la :: LocatedAn ann1 a2 -> LocatedAn ann2 a2

Helper function (temporary) during transition of names Discards any annotations

reLoc :: LocatedAn a e -> Located e

reLocA :: Located e -> LocatedAn ann e

reLocL :: LocatedN e -> LocatedA e

reLocC :: LocatedN e -> LocatedC e

reLocN :: LocatedN a -> Located a

la2r :: SrcSpanAnn' a -> RealSrcSpan

realSrcSpan :: SrcSpan -> RealSrcSpan

Building up annotations

extraToAnnList :: AnnList -> [AddEpAnn] -> AnnList

reAnn :: [TrailingAnn] -> EpAnnComments -> Located a -> LocatedA a

reAnnL :: ann -> EpAnnComments -> Located e -> GenLocated (SrcAnn ann) e

reAnnC :: AnnContext -> EpAnnComments -> Located a -> LocatedC a

addAnns :: EpAnn [AddEpAnn] -> [AddEpAnn] -> EpAnnComments -> EpAnn [AddEpAnn]

addAnnsA :: SrcSpanAnnA -> [TrailingAnn] -> EpAnnComments -> SrcSpanAnnA

widenSpan :: SrcSpan -> [AddEpAnn] -> SrcSpan

The annotations need to all come after the anchor. Make sure this is the case.

widenAnchor :: Anchor -> [AddEpAnn] -> Anchor

widenAnchorR :: Anchor -> RealSrcSpan -> Anchor

widenLocatedAn :: SrcSpanAnn' an -> [AddEpAnn] -> SrcSpanAnn' an

Querying annotations

getLocAnn :: Located a -> SrcSpanAnnA

epAnnAnns :: EpAnn [AddEpAnn] -> [AddEpAnn]

epAnnAnnsL :: EpAnn a -> [a]

annParen2AddEpAnn :: EpAnn AnnParen -> [AddEpAnn]

epAnnComments :: EpAnn an -> EpAnnComments

Working with locations of annotations

sortLocatedA :: [GenLocated (SrcSpanAnn' a) e] -> [GenLocated (SrcSpanAnn' a) e]

mapLocA :: (a -> b) -> GenLocated SrcSpan a -> GenLocated (SrcAnn ann) b

combineLocsA :: Semigroup a => GenLocated (SrcAnn a) e1 -> GenLocated (SrcAnn a) e2 -> SrcAnn a

combineSrcSpansA :: Semigroup a => SrcAnn a -> SrcAnn a -> SrcAnn a

addCLocA :: GenLocated (SrcSpanAnn' a) e1 -> GenLocated SrcSpan e2 -> e3 -> GenLocated (SrcAnn ann) e3

Combine locations from two Located things and add them to a third thing

addCLocAA :: GenLocated (SrcSpanAnn' a1) e1 -> GenLocated (SrcSpanAnn' a2) e2 -> e3 -> GenLocated (SrcAnn ann) e3

Constructing GenLocated annotation types when we do not care

noLocA :: a -> LocatedAn an a

getLocA :: GenLocated (SrcSpanAnn' a) e -> SrcSpan

noSrcSpanA :: SrcAnn ann

noAnnSrcSpan :: SrcSpan -> SrcAnn ann

Working with comments in annotations

noComments :: EpAnnCO

comment :: RealSrcSpan -> EpAnnComments -> EpAnnCO

addCommentsToSrcAnn :: Monoid ann => SrcAnn ann -> EpAnnComments -> SrcAnn ann

Add additional comments to a SrcAnn, used for manipulating the AST prior to exact printing the changed one.

setCommentsSrcAnn :: Monoid ann => SrcAnn ann -> EpAnnComments -> SrcAnn ann

Replace any existing comments on a SrcAnn, used for manipulating the AST prior to exact printing the changed one.

addCommentsToEpAnn :: Monoid a => SrcSpan -> EpAnn a -> EpAnnComments -> EpAnn a

Add additional comments, used for manipulating the AST prior to exact printing the changed one.

setCommentsEpAnn :: Monoid a => SrcSpan -> EpAnn a -> EpAnnComments -> EpAnn a

Replace any existing comments, used for manipulating the AST prior to exact printing the changed one.

transferAnnsA :: SrcSpanAnnA -> SrcSpanAnnA -> (SrcSpanAnnA, SrcSpanAnnA)

Transfer comments and trailing items from the annotations in the first SrcSpanAnnA argument to those in the second.

commentsOnlyA :: Monoid ann => SrcAnn ann -> SrcAnn ann

Remove the exact print annotations payload, leaving only the anchor and comments.

removeCommentsA :: SrcAnn ann -> SrcAnn ann

Remove the comments, leaving the exact print annotations payload

placeholderRealSpan :: RealSrcSpan