GHC.Rename.Unbound

Documentation

mkUnboundName :: OccName -> Name

mkUnboundNameRdr :: RdrName -> Name

mkUnboundGRE :: OccName -> GlobalRdrElt

mkUnboundGRERdr :: RdrName -> GlobalRdrElt

isUnboundName :: Name -> Bool

reportUnboundName :: RdrName -> RnM Name

reportUnboundName' :: WhatLooking -> RdrName -> RnM Name

unknownNameSuggestions :: LocalRdrEnv -> WhatLooking -> RdrName -> RnM ([ImportError], [GhcHint])

Called from the typechecker (GHC.Tc.Errors) when we find an unbound variable

similarNameSuggestions :: LookingFor -> DynFlags -> GlobalRdrEnv -> LocalRdrEnv -> RdrName -> [SimilarName]

fieldSelectorSuggestions :: GlobalRdrEnv -> RdrName -> [GhcHint]

When the name is in scope as field whose selector has been suppressed by NoFieldSelectors, display a helpful message explaining this.

data WhatLooking

Constructors

WL_Anything

WL_Constructor

WL_RecField

WL_None

data WhereLooking

Constructors

WL_Anywhere

WL_Global

WL_LocalTop

WL_LocalOnly

data LookingFor

Constructors

LF

Fields

lf_which :: WhatLooking

lf_where :: WhereLooking

unboundName :: LookingFor -> RdrName -> RnM Name

unboundNameX :: LookingFor -> RdrName -> [GhcHint] -> RnM Name

unboundTermNameInTypes :: LookingFor -> RdrName -> RdrName -> RnM Name

data IsTermInTypes

Constructors

UnknownTermInTypes RdrName

TermInTypes RdrName

NoTermInTypes

notInScopeErr :: WhereLooking -> RdrName -> NotInScopeError

nameSpacesRelated

Arguments

:: DynFlags
to find out whether -XDataKinds is enabled

-> WhatLooking
What kind of name are we looking for

-> NameSpace
Name space of the original name

-> NameSpace
Name space of a name that might have been meant

-> Bool

termNameInType :: LookingFor -> RdrName -> RdrName -> [GhcHint] -> RnM Name