GHC.Settings

Description

Run-time settings

Documentation

data Settings

Constructors

Settings

Fields

sGhcNameVersion :: !GhcNameVersion

sFileSettings :: !FileSettings

sTargetPlatform :: Platform

sToolSettings :: !ToolSettings

sPlatformMisc :: !PlatformMisc

sRawSettings :: [(String, String)]

data ToolSettings

Settings for other executables GHC calls.

Probably should further split down by phase, or split between platform-specific and platform-agnostic.

Constructors

ToolSettings

Fields

toolSettings_ldSupportsCompactUnwind :: Bool

toolSettings_ldSupportsBuildId :: Bool

toolSettings_ldSupportsFilelist :: Bool

toolSettings_ldIsGnuLd :: Bool

toolSettings_ccSupportsNoPie :: Bool

toolSettings_pgm_L :: String

toolSettings_pgm_P :: (String, [Option])

toolSettings_pgm_F :: String

toolSettings_pgm_c :: String

toolSettings_pgm_a :: (String, [Option])

toolSettings_pgm_l :: (String, [Option])

toolSettings_pgm_lm :: (String, [Option])

toolSettings_pgm_dll :: (String, [Option])

toolSettings_pgm_T :: String

toolSettings_pgm_windres :: String

toolSettings_pgm_libtool :: String

toolSettings_pgm_ar :: String

toolSettings_pgm_otool :: String

toolSettings_pgm_install_name_tool :: String

toolSettings_pgm_ranlib :: String

toolSettings_pgm_lo :: (String, [Option])

LLVM: opt llvm optimiser

toolSettings_pgm_lc :: (String, [Option])

LLVM: llc static compiler

toolSettings_pgm_lcc :: (String, [Option])

LLVM: c compiler

toolSettings_pgm_i :: String

toolSettings_opt_L :: [String]

toolSettings_opt_P :: [String]

toolSettings_opt_P_fingerprint :: Fingerprint

cached Fingerprint of sOpt_P See Note [Repeated -optP hashing]

toolSettings_opt_F :: [String]

toolSettings_opt_c :: [String]

toolSettings_opt_cxx :: [String]

toolSettings_opt_a :: [String]

toolSettings_opt_l :: [String]

toolSettings_opt_lm :: [String]

toolSettings_opt_windres :: [String]

toolSettings_opt_lo :: [String]

LLVM: llvm optimiser

toolSettings_opt_lc :: [String]

LLVM: llc static compiler

toolSettings_opt_lcc :: [String]

LLVM: c compiler

toolSettings_opt_i :: [String]

iserv options

toolSettings_extraGccViaCFlags :: [String]

data FileSettings

Paths to various files and directories used by GHC, including those that provide more settings.

Constructors

FileSettings

Fields

fileSettings_ghcUsagePath :: FilePath

fileSettings_ghciUsagePath :: FilePath

fileSettings_toolDir :: Maybe FilePath

fileSettings_topDir :: FilePath

fileSettings_tmpDir :: String

fileSettings_globalPackageDatabase :: FilePath

data GhcNameVersion

Settings for what GHC this is.

Constructors

GhcNameVersion

Fields

ghcNameVersion_programName :: String

ghcNameVersion_projectVersion :: String

data Platform

Platform description

This is used to describe platforms so that we can generate code for them.

Constructors

Platform

Fields

platformArchOS :: !ArchOS

Architecture and OS

platformWordSize :: !PlatformWordSize

Word size

platformByteOrder :: !ByteOrder

Byte order (endianness)

platformUnregisterised :: !Bool

platformHasGnuNonexecStack :: !Bool

platformHasIdentDirective :: !Bool

platformHasSubsectionsViaSymbols :: !Bool

platformIsCrossCompiling :: !Bool

platformLeadingUnderscore :: !Bool

Symbols need underscore prefix

platformTablesNextToCode :: !Bool

Determines whether we will be compiling info tables that reside just before the entry code, or with an indirection to the entry code. See TABLES_NEXT_TO_CODE in includesrtsstorage/InfoTables.h.

platform_constants :: !(Maybe PlatformConstants)

Constants such as structure offsets, type sizes, etc.

data PlatformMisc

Platform-specific settings formerly hard-coded in Config.hs.

These should probably be all be triaged whether they can be computed from other settings or belong in another another place (like Platform above).

Constructors

PlatformMisc

Fields

platformMisc_targetPlatformString :: String

platformMisc_ghcWithInterpreter :: Bool

platformMisc_ghcWithSMP :: Bool

platformMisc_ghcRTSWays :: String

platformMisc_libFFI :: Bool

platformMisc_ghcRtsWithLibdw :: Bool

platformMisc_llvmTarget :: String

Accessors

dynLibSuffix :: GhcNameVersion -> String

Dynamic library suffix

sProgramName :: Settings -> String

sProjectVersion :: Settings -> String

sGhcUsagePath :: Settings -> FilePath

sGhciUsagePath :: Settings -> FilePath

sToolDir :: Settings -> Maybe FilePath

sTopDir :: Settings -> FilePath

sTmpDir :: Settings -> String

sGlobalPackageDatabasePath :: Settings -> FilePath

sLdSupportsCompactUnwind :: Settings -> Bool

sLdSupportsBuildId :: Settings -> Bool

sLdSupportsFilelist :: Settings -> Bool

sLdIsGnuLd :: Settings -> Bool

sGccSupportsNoPie :: Settings -> Bool

sPgm_L :: Settings -> String

sPgm_P :: Settings -> (String, [Option])

sPgm_F :: Settings -> String

sPgm_c :: Settings -> String

sPgm_a :: Settings -> (String, [Option])

sPgm_l :: Settings -> (String, [Option])

sPgm_lm :: Settings -> (String, [Option])

sPgm_dll :: Settings -> (String, [Option])

sPgm_T :: Settings -> String

sPgm_windres :: Settings -> String

sPgm_libtool :: Settings -> String

sPgm_ar :: Settings -> String

sPgm_otool :: Settings -> String

sPgm_install_name_tool :: Settings -> String

sPgm_ranlib :: Settings -> String

sPgm_lo :: Settings -> (String, [Option])

sPgm_lc :: Settings -> (String, [Option])

sPgm_lcc :: Settings -> (String, [Option])

sPgm_i :: Settings -> String

sOpt_L :: Settings -> [String]

sOpt_P :: Settings -> [String]

sOpt_P_fingerprint :: Settings -> Fingerprint

sOpt_F :: Settings -> [String]

sOpt_c :: Settings -> [String]

sOpt_cxx :: Settings -> [String]

sOpt_a :: Settings -> [String]

sOpt_l :: Settings -> [String]

sOpt_lm :: Settings -> [String]

sOpt_windres :: Settings -> [String]

sOpt_lo :: Settings -> [String]

sOpt_lc :: Settings -> [String]

sOpt_lcc :: Settings -> [String]

sOpt_i :: Settings -> [String]

sExtraGccViaCFlags :: Settings -> [String]

sTargetPlatformString :: Settings -> String

sGhcWithInterpreter :: Settings -> Bool

sGhcWithSMP :: Settings -> Bool

sGhcRTSWays :: Settings -> String

sLibFFI :: Settings -> Bool

sGhcRtsWithLibdw :: Settings -> Bool