GHC.Stg.Utils

Documentation

mkStgAltTypeFromStgAlts :: forall p. Id -> [GenStgAlt p] -> AltType

bindersOf :: BinderP a ~ Id => GenStgBinding a -> [Id]

bindersOfX :: GenStgBinding a -> [BinderP a]

bindersOfTop :: BinderP a ~ Id => GenStgTopBinding a -> [Id]

bindersOfTopBinds :: BinderP a ~ Id => [GenStgTopBinding a] -> [Id]

stripStgTicksTop :: (StgTickish -> Bool) -> GenStgExpr p -> ([StgTickish], GenStgExpr p)

Strip ticks of a given type from an STG expression.

stripStgTicksTopE :: (StgTickish -> Bool) -> GenStgExpr p -> GenStgExpr p

Strip ticks of a given type from an STG expression returning only the expression.

idArgs :: [StgArg] -> [Id]

mkUnarisedId :: MonadUnique m => FastString -> UnaryType -> m Id

mkUnarisedIds :: MonadUnique m => FastString -> [UnaryType] -> m [Id]