GHC.StgToCmm.Utils

Documentation

emitDataLits :: CLabel -> [CmmLit] -> FCode ()

Emit a data-segment data block

emitRODataLits :: CLabel -> [CmmLit] -> FCode ()

Emit a read-only data block

emitDataCon :: CLabel -> CmmInfoTable -> CostCentreStack -> [CmmLit] -> FCode ()

emitRtsCall :: UnitId -> FastString -> [(CmmExpr, ForeignHint)] -> Bool -> FCode ()

emitRtsCallWithResult :: LocalReg -> ForeignHint -> UnitId -> FastString -> [(CmmExpr, ForeignHint)] -> Bool -> FCode ()

emitRtsCallGen :: [(LocalReg, ForeignHint)] -> CLabel -> [(CmmExpr, ForeignHint)] -> Bool -> FCode ()

assignTemp :: CmmExpr -> FCode LocalReg

newUnboxedTupleRegs :: Type -> FCode ([LocalReg], [ForeignHint])

emitMultiAssign :: [LocalReg] -> [CmmExpr] -> FCode ()

emitCmmLitSwitch :: CmmExpr -> [(Literal, CmmAGraphScoped)] -> CmmAGraphScoped -> FCode ()

emitSwitch :: CmmExpr -> [(ConTagZ, CmmAGraphScoped)] -> Maybe CmmAGraphScoped -> ConTagZ -> ConTagZ -> FCode ()

tagToClosure :: Platform -> TyCon -> CmmExpr -> CmmExpr

mkTaggedObjectLoad :: Platform -> LocalReg -> LocalReg -> ByteOff -> DynTag -> CmmAGraph

callerSaves :: Platform -> GlobalReg -> Bool

Returns True if this global register is stored in a caller-saves machine register.

callerSaveVolatileRegs :: Platform -> (CmmAGraph, CmmAGraph)

get_GlobalReg_addr :: Platform -> GlobalReg -> CmmExpr

We map STG registers onto appropriate CmmExprs. Either they map to real machine registers or stored as offsets from BaseReg. Given a GlobalReg, get_GlobalReg_addr always produces the register table address for it.

callerSaveGlobalReg :: Platform -> GlobalReg -> CmmAGraph

callerRestoreGlobalReg :: Platform -> GlobalReg -> CmmAGraph

cmmAndWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmOrWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmNegate :: Platform -> CmmExpr -> CmmExpr

cmmEqWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmNeWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmUGtWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmSubWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmMulWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmAddWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmUShrWord :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmOffsetExprW :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmOffsetExprB :: Platform -> CmmExpr -> CmmExpr -> CmmExpr

cmmRegOffW :: Platform -> CmmReg -> WordOff -> CmmExpr

cmmRegOffB :: CmmReg -> ByteOff -> CmmExpr

cmmLabelOffW :: Platform -> CLabel -> WordOff -> CmmLit

cmmLabelOffB :: CLabel -> ByteOff -> CmmLit

cmmOffsetW :: Platform -> CmmExpr -> WordOff -> CmmExpr

cmmOffsetB :: Platform -> CmmExpr -> ByteOff -> CmmExpr

cmmOffsetLitW :: Platform -> CmmLit -> WordOff -> CmmLit

cmmOffsetLitB :: CmmLit -> ByteOff -> CmmLit

cmmLoadIndexW :: Platform -> CmmExpr -> Int -> CmmType -> CmmExpr

cmmConstrTag1 :: Platform -> CmmExpr -> CmmExpr

cmmUntag :: Platform -> CmmExpr -> CmmExpr

cmmIsTagged :: Platform -> CmmExpr -> CmmExpr

addToMem :: CmmType -> CmmExpr -> Int -> CmmAGraph

addToMem rep ptr n adds n to the integer pointed-to by ptr.

addToMemE :: CmmType -> CmmExpr -> CmmExpr -> CmmAGraph

addToMemE rep ptr n adds n to the integer pointed-to by ptr.

addToMemLblE :: CmmType -> CLabel -> CmmExpr -> CmmAGraph

addToMemLbl :: CmmType -> CLabel -> Int -> CmmAGraph

Update remembered set operations

whenUpdRemSetEnabled :: FCode a -> FCode ()

emitUpdRemSetPush

Arguments

:: CmmExpr
value of pointer which was overwritten

-> FCode ()

Emit code to add an entry to a now-overwritten pointer to the update remembered set.

emitUpdRemSetPushThunk

Arguments

:: CmmExpr
the thunk

-> FCode ()

convertInfoProvMap :: DynFlags -> [CmmInfoTable] -> Module -> InfoTableProvMap -> [InfoProvEnt]

Convert source information collected about identifiers in Debug to entries suitable for placing into the info table provenenance table.

cmmInfoTableToInfoProvEnt :: Module -> CmmInfoTable -> InfoProvEnt

A bare bones InfoProvEnt for things which don't have a good source location