GHC.StgToJS.Linker.Types

Description

Documentation

data GhcjsEnv

Constructors

GhcjsEnv

Fields

linkerArchiveDeps :: MVar (Map (Set FilePath) (Map Module (Deps, DepsLocation), [LinkableUnit]))

newGhcjsEnv :: IO GhcjsEnv

return a fresh GhcjsEnv

data JSLinkConfig

Constructors

JSLinkConfig

Fields

lcNoJSExecutables :: Bool

lcNoHsMain :: Bool

lcOnlyOut :: Bool

lcNoRts :: Bool

lcNoStats :: Bool

defaultJSLinkConfig :: JSLinkConfig

generateAllJs :: JSLinkConfig -> Bool

we generate a runnable all.js only if we link a complete application, no incremental linking and no skipped parts

data LinkedObj

An object file that's either already in memory (with name) or on disk

Constructors

ObjFile FilePath
load from this file

ObjLoaded String Object
already loaded: description and payload

type LinkableUnit = (Module, Int)

A LinkableUnit is a pair of a module and the index of the block in the object file