GHC.StgToJS.Object

Description

Documentation

putObject

Arguments

:: BinHandle

-> ModuleName
module

-> Deps
dependencies

-> [ObjUnit]
linkable units and their symbols

-> IO ()

Given a handle to a Binary payload, add the module, mod_name, its dependencies, deps, and its linkable units to the payload.

getObjectHeader :: BinHandle -> IO (Either String ModuleName)

Parse object header

getObjectBody :: BinHandle -> ModuleName -> IO Object

Parse object body. Must be called after a sucessful getObjectHeader

getObject :: BinHandle -> IO (Maybe Object)

Parse object

readObject :: FilePath -> IO (Maybe Object)

Read object from file

The object is still in memory after this (see objHandle).

getObjectUnits :: Object -> (Word -> IndexEntry -> Bool) -> IO [ObjUnit]

Get units in the object file, using the given filtering function

readObjectUnits :: FilePath -> (Word -> IndexEntry -> Bool) -> IO [ObjUnit]

Read units in the object file, using the given filtering function

readObjectDeps :: FilePath -> IO (Maybe Deps)

Reads only the part necessary to get the dependencies

isGlobalUnit :: Int -> Bool

we use the convention that the first unit (0) is a module-global unit that's always included when something from the module is loaded. everything in a module implicitly depends on the global block. the global unit itself can't have dependencies

isJsObjectFile :: FilePath -> IO Bool

Test if the object file is a JS object

data Object

An object file

Constructors

Object

Fields

objModuleName :: !ModuleName

name of the module

objHandle :: !BinHandle

BinHandle that can be used to read the ObjUnits

objPayloadOffset :: !(Bin ObjUnit)

Offset of the payload (units)

objDeps :: !Deps

Dependencies

objIndex :: !Index

The Index, serialed unit indices and their linkable units

data IndexEntry

Constructors

IndexEntry

Fields

idxSymbols :: ![FastString]

Symbols exported by a unit

idxOffset :: !(Bin ObjUnit)

Offset of the unit in the object file

data Deps

dependencies for a single module

Constructors

Deps

Fields

depsModule :: !Module

module

depsRequired :: !BlockIds

blocks that always need to be linked when this object is loaded (e.g. everything that contains initializer code or foreign exports)

depsHaskellExported :: !(Map ExportedFun BlockId)

exported Haskell functions -> block

depsBlocks :: !(Array BlockId BlockDeps)

info about each block

data BlockDeps

Constructors

BlockDeps

Fields

blockBlockDeps :: [Int]

dependencies on blocks in this object

blockFunDeps :: [ExportedFun]

dependencies on exported symbols in other objects , blockForeignExported :: [ExpFun] , blockForeignImported :: [ForeignRef]

data DepsLocation

Where are the dependencies

Constructors

ObjectFile FilePath
In an object file at path

ArchiveFile FilePath
In a Ar file at path

InMemory String Object
In memory

data ExportedFun

Exported Functions

Constructors

ExportedFun

Fields

funModule :: !Module

The module containing the function

funSymbol :: !LexicalFastString

The function

Orphan instances

Binary Ident

Instance details

Methods

put_ :: BinHandle -> Ident -> IO ()

put :: BinHandle -> Ident -> IO (Bin Ident)

get :: BinHandle -> IO Ident

Binary JExpr

Instance details

Methods

put_ :: BinHandle -> JExpr -> IO ()

put :: BinHandle -> JExpr -> IO (Bin JExpr)

get :: BinHandle -> IO JExpr

Binary JOp

Instance details

Methods

put_ :: BinHandle -> JOp -> IO ()

put :: BinHandle -> JOp -> IO (Bin JOp)

get :: BinHandle -> IO JOp

Binary JStat

Instance details

Methods

put_ :: BinHandle -> JStat -> IO ()

put :: BinHandle -> JStat -> IO (Bin JStat)

get :: BinHandle -> IO JStat

Binary JUOp

Instance details

Methods

put_ :: BinHandle -> JUOp -> IO ()

put :: BinHandle -> JUOp -> IO (Bin JUOp)

get :: BinHandle -> IO JUOp

Binary JVal

Instance details

Methods

put_ :: BinHandle -> JVal -> IO ()

put :: BinHandle -> JVal -> IO (Bin JVal)

get :: BinHandle -> IO JVal

Binary SaneDouble

Instance details

Methods

put_ :: BinHandle -> SaneDouble -> IO ()

put :: BinHandle -> SaneDouble -> IO (Bin SaneDouble)

get :: BinHandle -> IO SaneDouble

Binary CILayout

Instance details

Methods

put_ :: BinHandle -> CILayout -> IO ()

put :: BinHandle -> CILayout -> IO (Bin CILayout)

get :: BinHandle -> IO CILayout

Binary CIRegs

Instance details

Methods

put_ :: BinHandle -> CIRegs -> IO ()

put :: BinHandle -> CIRegs -> IO (Bin CIRegs)

get :: BinHandle -> IO CIRegs

Binary CIStatic

Instance details

Methods

put_ :: BinHandle -> CIStatic -> IO ()

put :: BinHandle -> CIStatic -> IO (Bin CIStatic)

get :: BinHandle -> IO CIStatic

Binary CIType

Instance details

Methods

put_ :: BinHandle -> CIType -> IO ()

put :: BinHandle -> CIType -> IO (Bin CIType)

get :: BinHandle -> IO CIType

Binary ClosureInfo

Instance details

Methods

put_ :: BinHandle -> ClosureInfo -> IO ()

put :: BinHandle -> ClosureInfo -> IO (Bin ClosureInfo)

get :: BinHandle -> IO ClosureInfo

Binary ExpFun

Instance details

Methods

put_ :: BinHandle -> ExpFun -> IO ()

put :: BinHandle -> ExpFun -> IO (Bin ExpFun)

get :: BinHandle -> IO ExpFun

Binary ForeignJSRef

Instance details

Methods

put_ :: BinHandle -> ForeignJSRef -> IO ()

put :: BinHandle -> ForeignJSRef -> IO (Bin ForeignJSRef)

get :: BinHandle -> IO ForeignJSRef

Binary JSFFIType

Instance details

Methods

put_ :: BinHandle -> JSFFIType -> IO ()

put :: BinHandle -> JSFFIType -> IO (Bin JSFFIType)

get :: BinHandle -> IO JSFFIType

Binary StaticArg

Instance details

Methods

put_ :: BinHandle -> StaticArg -> IO ()

put :: BinHandle -> StaticArg -> IO (Bin StaticArg)

get :: BinHandle -> IO StaticArg

Binary StaticInfo

Instance details

Methods

put_ :: BinHandle -> StaticInfo -> IO ()

put :: BinHandle -> StaticInfo -> IO (Bin StaticInfo)

get :: BinHandle -> IO StaticInfo

Binary StaticLit

Instance details

Methods

put_ :: BinHandle -> StaticLit -> IO ()

put :: BinHandle -> StaticLit -> IO (Bin StaticLit)

get :: BinHandle -> IO StaticLit

Binary StaticUnboxed

Instance details

Methods

put_ :: BinHandle -> StaticUnboxed -> IO ()

put :: BinHandle -> StaticUnboxed -> IO (Bin StaticUnboxed)

get :: BinHandle -> IO StaticUnboxed

Binary StaticVal

Instance details

Methods

put_ :: BinHandle -> StaticVal -> IO ()

put :: BinHandle -> StaticVal -> IO (Bin StaticVal)

get :: BinHandle -> IO StaticVal

Binary VarType

Instance details

Methods

put_ :: BinHandle -> VarType -> IO ()

put :: BinHandle -> VarType -> IO (Bin VarType)

get :: BinHandle -> IO VarType