GHC.StgToJS.Symbols

Description

JS symbol generation

Documentation

moduleGlobalSymbol :: Module -> FastString

the global linkable unit of a module exports this symbol, depend on it to include that unit (used for cost centres)

moduleExportsSymbol :: Module -> FastString

mkJsSymbol :: Bool -> Module -> FastString -> FastString

Make JS symbol corresponding to the given Haskell symbol in the given module

mkJsSymbolBS :: Bool -> Module -> FastString -> ByteString

Make JS symbol corresponding to the given Haskell symbol in the given module

mkFreshJsSymbol :: Module -> Int -> FastString

Make JS symbol for given module and unique.

mkRawSymbol :: Bool -> FastString -> FastString

Make symbol "h$XYZ" or "h$$XYZ"

intBS :: Int -> ByteString

Hexadecimal representation of an int

Used for uniques. We could use base-62 as GHC usually does but this is likely faster.