GHC.StgToJS.Utils

Documentation

assignToTypedExprs :: [TypedExpr] -> [JExpr] -> JStat

assignCoerce1 :: [TypedExpr] -> [TypedExpr] -> JStat

Assign first expr only (if it exists), performing coercions between some PrimReps (e.g. StablePtr# and Addr#).

assignToExprCtx :: ExprCtx -> [JExpr] -> JStat

Core Utils

isUnboxableCon :: DataCon -> Bool

can we unbox C x to x, only if x is represented as a Number

isUnboxable :: VarType -> Bool

one-constructor types with one primitive field represented as a JS Number can be unboxed

data SlotCount

Number of slots occupied by a PrimRep

Constructors

NoSlot

OneSlot

TwoSlots

slotCount :: SlotCount -> Int

Return SlotCount as an Int

varSize :: VarType -> Int

Number of slots occupied by a value with the given VarType

varSlotCount :: VarType -> SlotCount

typeSize :: Type -> Int

isVoid :: VarType -> Bool

isPtr :: VarType -> Bool

isSingleVar :: VarType -> Bool

isMultiVar :: VarType -> Bool

isMatchable :: [VarType] -> Bool

can we pattern match on these values in a case?

tyConVt :: HasDebugCallStack => TyCon -> [VarType]

idVt :: HasDebugCallStack => Id -> [VarType]

typeVt :: HasDebugCallStack => Type -> [VarType]

uTypeVt :: HasDebugCallStack => UnaryType -> VarType

primRepVt :: HasDebugCallStack => PrimRep -> VarType

typePrimRep' :: HasDebugCallStack => UnaryType -> [PrimRep]

tyConPrimRep' :: HasDebugCallStack => TyCon -> [PrimRep]

Find the primitive representation of a TyCon. Defined here to avoid module loops. Call this only on unlifted tycons.

kindPrimRep' :: HasDebugCallStack => SDoc -> Kind -> [PrimRep]

Take a kind (of shape TYPE rr) and produce the PrimReps of values of types of this kind.

primTypeVt :: HasDebugCallStack => Type -> VarType

argVt :: StgArg -> VarType

dataConType :: DataCon -> Type

isBoolDataCon :: DataCon -> Bool

fixedLayout :: [VarType] -> CILayout

stackSlotType :: Id -> VarType

idPrimReps :: Id -> [PrimRep]

typePrimReps :: Type -> [PrimRep]

primRepSize :: PrimRep -> SlotCount

assocPrimReps :: [PrimRep] -> [JExpr] -> [(PrimRep, [JExpr])]

Associate the given values to each RrimRep in the given order, taking into account the number of slots per PrimRep

assocIdPrimReps :: Id -> [JExpr] -> [(PrimRep, [JExpr])]

Associate the given values to the Id's PrimReps, taking into account the number of slots per PrimRep

assocIdExprs :: Id -> [JExpr] -> [TypedExpr]

Associate the given JExpr to the Id's PrimReps, taking into account the number of slots per PrimRep

mkArityTag :: Int -> Int -> Int

toTypeList :: [VarType] -> [Int]

Stg Utils

bindingRefs :: UniqFM Id CgStgExpr -> CgStgBinding -> Set Id

collect Ids that this binding refers to (does not include the bindees themselves) first argument is Id -> StgExpr map for unfloated arguments

rhsRefs :: UniqFM Id CgStgExpr -> CgStgRhs -> Set Id

exprRefs :: UniqFM Id CgStgExpr -> CgStgExpr -> Set Id

altRefs :: UniqFM Id CgStgExpr -> CgStgAlt -> Set Id

argRefs :: UniqFM Id CgStgExpr -> StgArg -> Set Id

hasExport :: CgStgBinding -> Bool

collectTopIds :: CgStgBinding -> [Id]

collectIds :: UniqFM Id CgStgExpr -> CgStgBinding -> [Id]

removeTick :: CgStgExpr -> CgStgExpr

type LiveVars = DVarSet

liveStatic :: LiveVars -> LiveVars

liveVars :: LiveVars -> LiveVars

stgTopBindLive :: CgStgTopBinding -> [(Id, LiveVars)]

stgBindLive :: CgStgBinding -> [(Id, LiveVars)]

stgBindRhsLive :: CgStgBinding -> LiveVars

stgRhsLive :: CgStgRhs -> LiveVars

stgArgLive :: StgArg -> LiveVars

stgExprLive :: Bool -> CgStgExpr -> LiveVars

stgAltLive :: CgStgAlt -> LiveVars

stgLetNoEscapeLive :: Bool -> StgBinding -> StgExpr -> LiveVars

bindees :: CgStgBinding -> [Id]

isUpdatableRhs :: CgStgRhs -> Bool

stgLneLive :: CgStgBinding -> [Id]

stgLneLive' :: CgStgBinding -> [Id]

stgLneLiveExpr :: CgStgRhs -> [Id]

isInlineExpr :: UniqSet Id -> CgStgExpr -> (UniqSet Id, Bool)

returns True if the expression is definitely inline

inspectInlineBinding :: UniqSet Id -> CgStgBinding -> UniqSet Id

inspectInlineRhs :: UniqSet Id -> Id -> CgStgRhs -> UniqSet Id

isInlineForeignCall :: ForeignCall -> Bool

isInlineApp :: UniqSet Id -> Id -> [StgArg] -> Bool