GHC.SysTools.Tasks

Documentation

runUnlit :: Logger -> DynFlags -> [Option] -> IO ()

runCpp :: Logger -> DynFlags -> [Option] -> IO ()

runPp :: Logger -> DynFlags -> [Option] -> IO ()

runCc :: Maybe ForeignSrcLang -> Logger -> TmpFs -> DynFlags -> [Option] -> IO ()

Run compiler of C-like languages and raw objects (such as gcc or clang).

isContainedIn :: String -> String -> Bool

askLd :: Logger -> DynFlags -> [Option] -> IO String

Run the linker with some arguments and return the output

runAs :: Logger -> DynFlags -> [Option] -> IO ()

runLlvmOpt :: Logger -> DynFlags -> [Option] -> IO ()

Run the LLVM Optimiser

runLlvmLlc :: Logger -> DynFlags -> [Option] -> IO ()

Run the LLVM Compiler

runClang :: Logger -> DynFlags -> [Option] -> IO ()

Run the clang compiler (used as an assembler for the LLVM backend on OS X as LLVM doesn't support the OS X system assembler)

figureLlvmVersion :: Logger -> DynFlags -> IO (Maybe LlvmVersion)

Figure out which version of LLVM we are running this session

runLink :: Logger -> TmpFs -> DynFlags -> [Option] -> IO ()

runMergeObjects :: Logger -> TmpFs -> DynFlags -> [Option] -> IO ()

runLibtool :: Logger -> DynFlags -> [Option] -> IO ()

runAr :: Logger -> DynFlags -> Maybe FilePath -> [Option] -> IO ()

askOtool :: Logger -> DynFlags -> Maybe FilePath -> [Option] -> IO String

runInstallNameTool :: Logger -> DynFlags -> [Option] -> IO ()

runRanlib :: Logger -> DynFlags -> [Option] -> IO ()

runWindres :: Logger -> DynFlags -> [Option] -> IO ()

touch :: Logger -> DynFlags -> String -> String -> IO ()

Tracing utility

traceToolCommand :: Logger -> DynFlags -> String -> IO a -> IO a

Record in the eventlog when the given tool command starts and finishes, prepending the given String with "systool:", to easily be able to collect and process all the systool events.

For those events to show up in the eventlog, you need to run GHC with -v2 or -ddump-timings.