GHC.Tc.Deriv.Generate

Description

Generating derived instance declarations

This module is nominally `subordinate' to GHC.Tc.Deriv, which is the `official' interface to deriving-related things.

This is where we do all the grimy bindings' generation.

Documentation

data AuxBindSpec

A declarative description of an auxiliary binding that should be generated. See Note [Auxiliary binders] for a more detailed description of how these are used.

Constructors

DerivTag2Con TyCon RdrName
$tag2con: Given a tag, computes the corresponding data constructor

DerivMaxTag TyCon RdrName
$maxtag: The maximum possible tag value among a data type's constructors

DerivDataDataType TyCon RdrName [RdrName]
$t: The DataType representation for a Data instance

DerivDataConstr DataCon RdrName RdrName
$c: The Constr representation for a Data instance

gen_Eq_binds :: SrcSpan -> DerivInstTys -> TcM (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Ord_binds :: SrcSpan -> DerivInstTys -> TcM (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Enum_binds :: SrcSpan -> DerivInstTys -> TcM (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Bounded_binds :: SrcSpan -> DerivInstTys -> (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Ix_binds :: SrcSpan -> DerivInstTys -> TcM (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Show_binds :: (Name -> Fixity) -> SrcSpan -> DerivInstTys -> (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Read_binds :: (Name -> Fixity) -> SrcSpan -> DerivInstTys -> (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Data_binds :: SrcSpan -> DerivInstTys -> TcM (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Lift_binds :: SrcSpan -> DerivInstTys -> (LHsBinds GhcPs, Bag AuxBindSpec)

gen_Newtype_binds :: SrcSpan -> Class -> [TyVar] -> [Type] -> Type -> (LHsBinds GhcPs, [LSig GhcPs])

gen_Newtype_fam_insts :: SrcSpan -> Class -> [TyVar] -> [Type] -> Type -> TcM [FamInst]

mkCoerceClassMethEqn :: Class -> [TyVar] -> [Type] -> Type -> Id -> Pair Type

genAuxBinds :: SrcSpan -> Bag AuxBindSpec -> Bag (LHsBind GhcPs, LSig GhcPs)

Take a Bag of AuxBindSpecs and generate the code for auxiliary bindings based on the declarative descriptions in the supplied AuxBindSpecs. See Note [Auxiliary binders].

ordOpTbl :: [(Type, (RdrName, RdrName, RdrName, RdrName, RdrName))]

boxConTbl :: [(Type, LHsExpr GhcPs -> LHsExpr GhcPs)]

litConTbl :: [(Type, LHsExpr GhcPs -> LHsExpr GhcPs)]

mkRdrFunBind :: LocatedN RdrName -> [LMatch GhcPs (LHsExpr GhcPs)] -> LHsBind GhcPs

mkRdrFunBindEC :: Arity -> (LHsExpr GhcPs -> LHsExpr GhcPs) -> LocatedN RdrName -> [LMatch GhcPs (LHsExpr GhcPs)] -> LHsBind GhcPs

Produces a function binding. When no equations are given, it generates a binding of the given arity and an empty case expression for the last argument that it passes to the given function to produce the right-hand side.

mkRdrFunBindSE :: Arity -> LocatedN RdrName -> [LMatch GhcPs (LHsExpr GhcPs)] -> LHsBind GhcPs

Produces a function binding. When there are no equations, it generates a binding with the given arity that produces an error based on the name of the type of the last argument.

error_Expr :: FastString -> LHsExpr GhcPs

getPossibleDataCons :: TyCon -> [Type] -> [DataCon]

getPossibleDataCons tycon tycon_args returns the constructors of tycon whose return types match when checked against tycon_args.

See Note [Filter out impossible GADT data constructors]

data DerivInstTys

Information about the arguments to the class in a stock- or newtype-derived instance. For a deriving-generated instance declaration such as this one:

instance Ctx => Cls cls_ty_1 ... cls_ty_m (TC tc_arg_1 ... tc_arg_n) where ...

dit_cls_tys corresponds to cls_ty_1 ... cls_ty_m.

dit_tc corresponds to TC.

dit_tc_args corresponds to tc_arg_1 ... tc_arg_n.

See Note [DerivEnv and DerivSpecMechanism] in GHC.Tc.Deriv.Utils for a more in-depth explanation, including the relationship between dit_tcdit_rep_tc and dit_tc_argsdit_rep_tc_args.

A DerivInstTys value can be seen as a more structured representation of the denv_inst_tys in a DerivEnv, as the denv_inst_tys is equal to dit_cls_tys ++ [mkTyConApp dit_tc dit_tc_args]. Other parts of the instance declaration can be found in the DerivEnv. For example, the Cls in the example above corresponds to the denv_cls field of DerivEnv.

Similarly, the type variables that appear in a DerivInstTys value are the same type variables as the denv_tvs in the parent DerivEnv. Accordingly, if we are inferring an instance context, the type variables will be TcTyVar skolems. Otherwise, they will be ordinary TyVars. See Note [Overlap and deriving] in GHC.Tc.Deriv.Infer.

Constructors

DerivInstTys

Fields

dit_cls_tys :: [Type]

Other arguments to the class except the last

dit_tc :: TyCon

Type constructor for which the instance is requested (last arguments to the type class)

dit_tc_args :: [Type]

Arguments to the type constructor

dit_rep_tc :: TyCon

The representation tycon for dit_tc (for data family instances). Otherwise the same as dit_tc.

dit_rep_tc_args :: [Type]

The representation types for dit_tc_args (for data family instances). Otherwise the same as dit_tc_args.

dit_dc_inst_arg_env :: DataConEnv [Type]

The cached results of instantiating each data constructor's field types using dataConInstUnivs data_con dit_rep_tc_args. See Note [Instantiating field types in stock deriving].

This field is only used for stock-derived instances and goes unused for newtype-derived instances. It is put here mainly for the sake of convenience.

buildDataConInstArgEnv :: TyCon -> [Type] -> DataConEnv [Type]

buildDataConInstArgEnv tycon arg_tys constructs a cache that maps each of tycon's data constructors to their field types, with are to be instantiated with arg_tys. See Note [Instantiating field types in stock deriving].

derivDataConInstArgTys :: DataCon -> DerivInstTys -> [Type]

Look up a data constructor's instantiated field types in a DerivInstTys. See Note [Instantiating field types in stock deriving].

substDerivInstTys :: Subst -> DerivInstTys -> DerivInstTys

Apply a substitution to all of the Types contained in a DerivInstTys. See Note [Instantiating field types in stock deriving] for why we need to substitute into a DerivInstTys in the first place.

zonkDerivInstTys :: ZonkEnv -> DerivInstTys -> TcM DerivInstTys

Zonk the TcTyVars in a DerivInstTys value to TyVars. See Note [What is zonking?] in GHC.Tc.Utils.TcMType.

This is only used in the final zonking step when inferring the context for a derived instance. See Note [Overlap and deriving] in GHC.Tc.Deriv.Infer.