GHC.Tc.Solver.InertSet

The work list

data WorkList

Constructors

WL

Fields

wl_eqs :: [Ct]

wl_rest :: [Ct]

wl_implics :: Bag Implication

isEmptyWorkList :: WorkList -> Bool

emptyWorkList :: WorkList

extendWorkListNonEq :: Ct -> WorkList -> WorkList

extendWorkListCt :: Ct -> WorkList -> WorkList

extendWorkListCts :: [Ct] -> WorkList -> WorkList

extendWorkListEq :: Ct -> WorkList -> WorkList

appendWorkList :: WorkList -> WorkList -> WorkList

extendWorkListImplic :: Implication -> WorkList -> WorkList

workListSize :: WorkList -> Int

selectWorkItem :: WorkList -> Maybe (Ct, WorkList)

The inert set

data InertSet

Constructors

IS

Fields

inert_cans :: InertCans

inert_cycle_breakers :: CycleBreakerVarStack

inert_famapp_cache :: FunEqMap Reduction

inert_solved_dicts :: DictMap CtEvidence

data InertCans

Constructors

IC

Fields

inert_eqs :: InertEqs

inert_funeqs :: FunEqMap EqualCtList

inert_dicts :: DictMap Ct

inert_insts :: [QCInst]

inert_safehask :: DictMap Ct

See Note [Safe Haskell Overlapping Instances Implementation] in GHC.Tc.Solver

inert_irreds :: Cts

inert_given_eq_lvl :: TcLevel

inert_given_eqs :: Bool

type InertEqs = DTyVarEnv EqualCtList

emptyInert :: InertSet

addInertItem :: TcLevel -> InertCans -> Ct -> InertCans

noMatchableGivenDicts :: InertSet -> CtLoc -> Class -> [TcType] -> Bool

Returns True iff there are no Given constraints that might, potentially, match the given class consraint. This is used when checking to see if a Given might overlap with an instance. See Note [Instance and Given overlap] in GHC.Tc.Solver.Interact

noGivenNewtypeReprEqs :: TyCon -> InertSet -> Bool

mightEqualLater :: InertSet -> TcPredType -> CtLoc -> TcPredType -> CtLoc -> Maybe Subst

prohibitedSuperClassSolve

Arguments

:: CtLoc
is it loopy to use this one ...

-> CtLoc
... to solve this one?

-> Bool
True ==> don't solve it

Is it (potentially) loopy to use the first ct1 to solve ct2?

Necessary (but not sufficient) conditions for this function to return True:

ct1 and ct2 both arise from superclass expansion,

ct1 is a Given and ct2 is a Wanted.

See Note [Solving superclass constraints] in GHC.Tc.TyCl.Instance, (sc2).

Inert equalities

foldTyEqs :: (Ct -> b -> b) -> InertEqs -> b -> b

delEq :: InertCans -> CanEqLHS -> TcType -> InertCans

findEq :: InertCans -> CanEqLHS -> [Ct]

partitionInertEqs :: (Ct -> Bool) -> InertEqs -> ([Ct], InertEqs)

partitionFunEqs :: (Ct -> Bool) -> FunEqMap EqualCtList -> ([Ct], FunEqMap EqualCtList)

Kick-out

kickOutRewritableLHS :: CtFlavourRole -> CanEqLHS -> InertCans -> (WorkList, InertCans)

Cycle breaker vars

type CycleBreakerVarStack

Arguments

= NonEmpty [(TcTyVar, TcType)]
a stack of (CycleBreakerTv, original family applications) lists first element in the stack corresponds to current implication; later elements correspond to outer implications used to undo the cycle-breaking needed to handle Note [Type equality cycles] in GHC.Tc.Solver.Canonical Why store the outer implications? For the use in mightEqualLater (only)

pushCycleBreakerVarStack :: CycleBreakerVarStack -> CycleBreakerVarStack

Push a fresh environment onto the cycle-breaker var stack. Useful when entering a nested implication.

insertCycleBreakerBinding

Arguments

:: TcTyVar
cbv, must be a CycleBreakerTv

-> TcType
cbv's expansion

-> CycleBreakerVarStack

-> CycleBreakerVarStack

Add a new cycle-breaker binding to the top environment on the stack.

forAllCycleBreakerBindings_ :: Monad m => CycleBreakerVarStack -> (TcTyVar -> TcType -> m ()) -> m ()

Perform a monadic operation on all pairs in the top environment in the stack.