GHC.Tc.Solver.Monad

Description

Monadic definitions for the constraint solver

Documentation

data TcS a

runTcS :: TcS a -> TcM (a, EvBindMap)

runTcSEarlyAbort :: TcS a -> TcM a

This variant of runTcS will immediately fail upon encountering an insoluble ct. See Note [Speeding up valid hole-fits]. Its one usage site does not need the ev_binds, so we do not return them.

runTcSWithEvBinds :: EvBindsVar -> TcS a -> TcM a

runTcSInerts :: InertSet -> TcS a -> TcM (a, InertSet)

A variant of runTcS that takes and returns an InertSet for later resumption of the TcS session.

failTcS :: TcRnMessage -> TcS a

warnTcS :: TcRnMessage -> TcS ()

addErrTcS :: TcRnMessage -> TcS ()

wrapTcS :: TcM a -> TcS a

ctLocWarnTcS :: CtLoc -> TcRnMessage -> TcS ()

Emit a warning within the TcS monad at the location given by the CtLoc.

runTcSEqualities :: TcS a -> TcM a

This can deal only with equality constraints.

nestTcS :: TcS a -> TcS a

nestImplicTcS :: EvBindsVar -> TcLevel -> TcS a -> TcS a

setEvBindsTcS :: EvBindsVar -> TcS a -> TcS a

emitImplicationTcS :: TcLevel -> SkolemInfoAnon -> [TcTyVar] -> [EvVar] -> Cts -> TcS TcEvBinds

emitTvImplicationTcS :: TcLevel -> SkolemInfoAnon -> [TcTyVar] -> Cts -> TcS ()

selectNextWorkItem :: TcS (Maybe Ct)

getWorkList :: TcS WorkList

updWorkListTcS :: (WorkList -> WorkList) -> TcS ()

pushLevelNoWorkList :: SDoc -> TcS a -> TcS (TcLevel, a)

runTcPluginTcS :: TcPluginM a -> TcS a

recordUsedGREs :: Bag GlobalRdrElt -> TcS ()

matchGlobalInst :: DynFlags -> Bool -> Class -> [Type] -> TcS ClsInstResult

data ClsInstResult

Constructors

NoInstance

OneInst

Fields

cir_new_theta :: [TcPredType]

cir_mk_ev :: [EvExpr] -> EvTerm

cir_what :: InstanceWhat

NotSure

data QCInst

Constructors

QCI

Fields

qci_ev :: CtEvidence

qci_tvs :: [TcTyVar]

qci_pred :: TcPredType

qci_pend_sc :: Bool

panicTcS :: SDoc -> TcS a

traceTcS :: String -> SDoc -> TcS ()

traceFireTcS :: CtEvidence -> SDoc -> TcS ()

bumpStepCountTcS :: TcS ()

csTraceTcS :: SDoc -> TcS ()

wrapErrTcS :: TcM a -> TcS a

wrapWarnTcS :: TcM a -> TcS a

resetUnificationFlag :: TcS Bool

setUnificationFlag :: TcLevel -> TcS ()

data MaybeNew

Constructors

Fresh CtEvidence

Cached EvExpr

freshGoals :: [MaybeNew] -> [CtEvidence]

isFresh :: MaybeNew -> Bool

getEvExpr :: MaybeNew -> EvExpr

newTcEvBinds :: TcS EvBindsVar

newNoTcEvBinds :: TcS EvBindsVar

newWantedEq :: CtLoc -> RewriterSet -> Role -> TcType -> TcType -> TcS (CtEvidence, Coercion)

Create a new Wanted constraint holding a coercion hole for an equality between the two types at the given Role.

emitNewWantedEq :: CtLoc -> RewriterSet -> Role -> TcType -> TcType -> TcS Coercion

Emit a new Wanted equality into the work-list

newWanted :: CtLoc -> RewriterSet -> PredType -> TcS MaybeNew

Create a new Wanted constraint, potentially looking up non-equality constraints in the cache instead of creating a new one from scratch.

Deals with both equality and non-equality constraints.

newWantedNC :: CtLoc -> RewriterSet -> PredType -> TcS CtEvidence

Create a new Wanted constraint.

Deals with both equality and non-equality constraints.

Does not attempt to re-use non-equality constraints that already exist in the inert set.

newWantedEvVarNC :: CtLoc -> RewriterSet -> TcPredType -> TcS CtEvidence

Create a new Wanted constraint holding an evidence variable.

Don't use this for equality constraints: use newWantedEq instead.

newBoundEvVarId :: TcPredType -> EvTerm -> TcS EvVar

Make a new Id of the given type, bound (in the monad's EvBinds) to the given term

unifyTyVar :: TcTyVar -> TcType -> TcS ()

reportUnifications :: TcS a -> TcS (Int, a)

touchabilityTest :: CtFlavour -> TcTyVar -> TcType -> TcS (TouchabilityTestResult, TcType)

This is the key test for untouchability: See Note [Unification preconditions] in GHC.Tc.Utils.Unify and Note [Solve by unification] in GHC.Tc.Solver.Interact

Returns a new rhs type, as this function can turn make some metavariables concrete.

data TouchabilityTestResult

Constructors

TouchableSameLevel

TouchableOuterLevel [TcTyVar] TcLevel

Untouchable

setEvBind :: EvBind -> TcS ()

setWantedEq :: HasDebugCallStack => TcEvDest -> Coercion -> TcS ()

Equalities only

setWantedEvTerm :: TcEvDest -> EvTerm -> TcS ()

Good for both equalities and non-equalities

setEvBindIfWanted :: CtEvidence -> EvTerm -> TcS ()

newEvVar :: TcPredType -> TcS EvVar

newGivenEvVar :: CtLoc -> (TcPredType, EvTerm) -> TcS CtEvidence

newGivenEvVars :: CtLoc -> [(TcPredType, EvTerm)] -> TcS [CtEvidence]

checkReductionDepth

Arguments

:: CtLoc

-> TcType
type being reduced

-> TcS ()

Checks if the depth of the given location is too much. Fails if it's too big, with an appropriate error message.

getSolvedDicts :: TcS (DictMap CtEvidence)

setSolvedDicts :: DictMap CtEvidence -> TcS ()

getInstEnvs :: TcS InstEnvs

getFamInstEnvs :: TcS (FamInstEnv, FamInstEnv)

getTopEnv :: TcS HscEnv

getGblEnv :: TcS TcGblEnv

getLclEnv :: TcS TcLclEnv

setLclEnv :: TcLclEnv -> TcS a -> TcS a

getTcEvBindsVar :: TcS EvBindsVar

getTcLevel :: TcS TcLevel

getTcEvTyCoVars :: EvBindsVar -> TcS TyCoVarSet

getTcEvBindsMap :: EvBindsVar -> TcS EvBindMap

setTcEvBindsMap :: EvBindsVar -> EvBindMap -> TcS ()

tcLookupClass :: Name -> TcS Class

tcLookupId :: Name -> TcS Id

updInertTcS :: (InertSet -> InertSet) -> TcS ()

updInertCans :: (InertCans -> InertCans) -> TcS ()

updInertDicts :: (DictMap Ct -> DictMap Ct) -> TcS ()

updInertIrreds :: (Cts -> Cts) -> TcS ()

getHasGivenEqs :: TcLevel -> TcS (HasGivenEqs, Cts)

setInertCans :: InertCans -> TcS ()

getInertEqs :: TcS InertEqs

getInertCans :: TcS InertCans

getInertGivens :: TcS [Ct]

getInertInsols :: TcS Cts

getInnermostGivenEqLevel :: TcS TcLevel

getTcSInerts :: TcS InertSet

setTcSInerts :: InertSet -> TcS ()

getUnsolvedInerts :: TcS (Bag Implication, Cts)

removeInertCts :: [Ct] -> InertCans -> InertCans

Remove inert constraints from the InertCans, for use when a typechecker plugin wishes to discard a given.

getPendingGivenScs :: TcS [Ct]

addInertCan :: Ct -> TcS ()

insertFunEq :: FunEqMap a -> TyCon -> [Type] -> a -> FunEqMap a

addInertForAll :: QCInst -> TcS ()

emitWorkNC :: [CtEvidence] -> TcS ()

emitWork :: [Ct] -> TcS ()

lookupInertDict :: InertCans -> CtLoc -> Class -> [Type] -> Maybe Ct

Look up a dictionary inert.

kickOutAfterUnification :: TcTyVar -> TcS Int

addInertSafehask :: InertCans -> Ct -> InertCans

insertSafeOverlapFailureTcS :: InstanceWhat -> Ct -> TcS ()

updInertSafehask :: (DictMap Ct -> DictMap Ct) -> TcS ()

getSafeOverlapFailures :: TcS Cts

addSolvedDict :: InstanceWhat -> CtEvidence -> Class -> [Type] -> TcS ()

lookupSolvedDict :: InertSet -> CtLoc -> Class -> [Type] -> Maybe CtEvidence

Look up a solved inert.

foldIrreds :: (Ct -> b -> b) -> Cts -> b -> b

lookupFamAppInert :: (CtFlavourRole -> Bool) -> TyCon -> [Type] -> TcS (Maybe (Reduction, CtFlavourRole))

Looks up a family application in the inerts.

lookupFamAppCache :: TyCon -> [Type] -> TcS (Maybe Reduction)

extendFamAppCache :: TyCon -> [Type] -> Reduction -> TcS ()

pprKicked :: Int -> SDoc

instDFunType :: DFunId -> [DFunInstType] -> TcS ([TcType], TcThetaType)

newFlexiTcSTy :: Kind -> TcS TcType

instFlexiX :: Subst -> [TKVar] -> TcS Subst

cloneMetaTyVar :: TcTyVar -> TcS TcTyVar

tcInstSkolTyVarsX :: SkolemInfo -> Subst -> [TyVar] -> TcS (Subst, [TcTyVar])

data TcLevel

isFilledMetaTyVar_maybe :: TcTyVar -> TcS (Maybe Type)

isFilledMetaTyVar :: TcTyVar -> TcS Bool

zonkTyCoVarsAndFV :: TcTyCoVarSet -> TcS TcTyCoVarSet

zonkTcType :: TcType -> TcS TcType

zonkTcTypes :: [TcType] -> TcS [TcType]

zonkTcTyVar :: TcTyVar -> TcS TcType

zonkCo :: Coercion -> TcS Coercion

zonkTyCoVarsAndFVList :: [TcTyCoVar] -> TcS [TcTyCoVar]

zonkSimples :: Cts -> TcS Cts

zonkWC :: WantedConstraints -> TcS WantedConstraints

zonkTyCoVarKind :: TcTyCoVar -> TcS TcTyCoVar

newTcRef :: a -> TcS (TcRef a)

readTcRef :: TcRef a -> TcS a

writeTcRef :: TcRef a -> a -> TcS ()

updTcRef :: TcRef a -> (a -> a) -> TcS ()

getDefaultInfo :: TcS ([Type], (Bool, Bool))

getDynFlags :: HasDynFlags m => m DynFlags

getGlobalRdrEnvTcS :: TcS GlobalRdrEnv

matchFam :: TyCon -> [Type] -> TcS (Maybe ReductionN)

matchFamTcM :: TyCon -> [Type] -> TcM (Maybe ReductionN)

checkWellStagedDFun :: CtLoc -> InstanceWhat -> PredType -> TcS ()

pprEq :: TcType -> TcType -> SDoc

breakTyEqCycle_maybe :: CtEvidence -> CheckTyEqResult -> CanEqLHS -> TcType -> TcS (Maybe ReductionN)

Conditionally replace all type family applications in the RHS with fresh variables, emitting givens that relate the type family application to the variable. See Note [Type equality cycles] in GHC.Tc.Solver.Canonical. This only works under conditions as described in the Note; otherwise, returns Nothing.

rewriterView :: TcType -> Maybe TcType