GHC.Tc.Types.Origin

Description

Describes the provenance of types as they flow through the type-checker. The datatypes here are mainly used for error message generation.

Documentation

data UserTypeCtxt

UserTypeCtxt describes the origin of the polymorphic type in the places where we need an expression to have that type

Constructors

FunSigCtxt Name Bool

InfSigCtxt Name

ExprSigCtxt

KindSigCtxt

StandaloneKindSigCtxt Name

TypeAppCtxt

ConArgCtxt Name

TySynCtxt Name

PatSynCtxt Name

PatSigCtxt

RuleSigCtxt Name

ForSigCtxt Name

DefaultDeclCtxt

InstDeclCtxt Bool

SpecInstCtxt

GenSigCtxt

GhciCtxt Bool

ClassSCCtxt Name

SigmaCtxt

DataTyCtxt Name

DerivClauseCtxt

TyVarBndrKindCtxt Name

DataKindCtxt Name

TySynKindCtxt Name

TyFamResKindCtxt Name

pprUserTypeCtxt :: UserTypeCtxt -> SDoc

isSigMaybe :: UserTypeCtxt -> Maybe Name

data SkolemInfo

Constructors

SigSkol UserTypeCtxt TcType [(Name, TcTyVar)]

SigTypeSkol UserTypeCtxt

ForAllSkol SDoc

DerivSkol Type

InstSkol

FamInstSkol

PatSkol ConLike (HsMatchContext GhcRn)

ArrowSkol

IPSkol [HsIPName]

RuleSkol RuleName

InferSkol [(Name, TcType)]

BracketSkol

UnifyForAllSkol TcType

TyConSkol TyConFlavour Name

DataConSkol Name

ReifySkol

QuantCtxtSkol

RuntimeUnkSkol

UnkSkol

pprSigSkolInfo :: UserTypeCtxt -> TcType -> SDoc

pprSkolInfo :: SkolemInfo -> SDoc

data CtOrigin

Constructors

GivenOrigin SkolemInfo
A given constraint from a user-written type signature. The SkolemInfo inside gives more information.

InstSCOrigin
InstSCOrigin is used for a Given constraint obtained by superclass selection from the context of an instance declaration. E.g. instance (Foo a, Bar a) => C [a] where ... When typechecking the instance decl itself, including producing evidence for the superclasses of C, the superclasses of (Foo a) and (Bar a) will have InstSCOrigin origin.

Fields

ScDepth

The number of superclass selections necessary to get this constraint; see Note [Replacement vs keeping] and Note [Use only the best local instance], both in GHC.Tc.Solver.Interact

TypeSize

If (C ty1 .. tyn) is the largest class from which we made a superclass selection in the chain, then TypeSize = sizeTypes [ty1, .., tyn] See Note [Solving superclass constraints] in GHC.Tc.TyCl.Instance

OtherSCOrigin
OtherSCOrigin is used for a Given constraint obtained by superclass selection from a constraint other than the context of an instance declaration. (For the latter we use InstSCOrigin.) E.g. f :: Foo a => blah f = e When typechecking body of f, the superclasses of the Given (Foo a) will have OtherSCOrigin. Needed for Note [Replacement vs keeping] and Note [Use only the best local instance], both in GHC.Tc.Solver.Interact.

Fields

ScDepth

The number of superclass selections necessary to get this constraint

SkolemInfo

Where the sub-class constraint arose from (used only for printing)

OccurrenceOf Name

OccurrenceOfRecSel RdrName

AppOrigin

SpecPragOrigin UserTypeCtxt

TypeEqOrigin

Fields

uo_actual :: TcType

uo_expected :: TcType

uo_thing :: Maybe SDoc

The thing that has type "actual"

uo_visible :: Bool

Is at least one of the three elements above visible? (Errors from the polymorphic subsumption check are considered visible.) Only used for prioritizing error messages.

KindEqOrigin TcType TcType CtOrigin (Maybe TypeOrKind)

IPOccOrigin HsIPName

OverLabelOrigin FastString

LiteralOrigin (HsOverLit GhcRn)

NegateOrigin

ArithSeqOrigin (ArithSeqInfo GhcRn)

AssocFamPatOrigin

SectionOrigin

HasFieldOrigin FastString

TupleOrigin

ExprSigOrigin

PatSigOrigin

PatOrigin

ProvCtxtOrigin (PatSynBind GhcRn GhcRn)

RecordUpdOrigin

ViewPatOrigin

ScOrigin TypeSize
ScOrigin is used only for the Wanted constraints for the superclasses of an instance declaration. If the instance head is C ty1 .. tyn then TypeSize = sizeTypes [ty1, .., tyn] See Note [Solving superclass constraints] in GHC.Tc.TyCl.Instance

DerivClauseOrigin

DerivOriginDC DataCon Int Bool

DerivOriginCoerce Id Type Type Bool

StandAloneDerivOrigin

DefaultOrigin

DoOrigin

DoPatOrigin (LPat GhcRn)

MCompOrigin

MCompPatOrigin (LPat GhcRn)

IfOrigin

ProcOrigin

AnnOrigin

FunDepOrigin1 PredType CtOrigin RealSrcSpan PredType CtOrigin RealSrcSpan

FunDepOrigin2 PredType CtOrigin PredType SrcSpan

ExprHoleOrigin (Maybe OccName)

TypeHoleOrigin OccName

PatCheckOrigin

ListOrigin

BracketOrigin

StaticOrigin

Shouldn'tHappenOrigin String

InstProvidedOrigin Module ClsInst

NonLinearPatternOrigin

UsageEnvironmentOf Name

CycleBreakerOrigin CtOrigin

InstanceSigOrigin Name Type Type

AmbiguityCheckOrigin UserTypeCtxt

GhcBug20076

exprCtOrigin :: HsExpr GhcRn -> CtOrigin

lexprCtOrigin :: LHsExpr GhcRn -> CtOrigin

Extract a suitable CtOrigin from a HsExpr

matchesCtOrigin :: MatchGroup GhcRn (LHsExpr GhcRn) -> CtOrigin

Extract a suitable CtOrigin from a MatchGroup

grhssCtOrigin :: GRHSs GhcRn (LHsExpr GhcRn) -> CtOrigin

Extract a suitable CtOrigin from guarded RHSs

isVisibleOrigin :: CtOrigin -> Bool

toInvisibleOrigin :: CtOrigin -> CtOrigin

pprCtOrigin :: CtOrigin -> SDoc

isGivenOrigin :: CtOrigin -> Bool