GHC.Tc.Utils.TcType

Description

Types used in the typechecker

This module provides the Type interface for front-end parts of the compiler. These parts

treat "source types" as opaque: newtypes, and predicates are meaningful.

look through usage types

Documentation

type TcType = Type

type TcSigmaType = TcType

type TcRhoType = TcType

type TcTauType = TcType

type TcPredType = PredType

type TcThetaType = ThetaType

type TcTyVar = Var

Type variable that might be a metavariable

type TcTyVarSet = TyVarSet

type TcDTyVarSet = DTyVarSet

type TcTyCoVarSet = TyCoVarSet

type TcDTyCoVarSet = DTyCoVarSet

type TcKind = Kind

type TcCoVar = CoVar

type TcTyCoVar = Var

type TcTyVarBinder = TyVarBinder

type TcInvisTVBinder = InvisTVBinder

type TcReqTVBinder = ReqTVBinder

type TcTyCon = TyCon

type KnotTied ty = ty

A type labeled KnotTied might have knot-tied tycons in it. See Note [Type checking recursive type and class declarations] in GHC.Tc.TyCl

data ExpType

An expected type to check against during type-checking. See Note [ExpType] in GHC.Tc.Utils.TcMType, where you'll also find manipulators.

Constructors

Check TcType

Infer !InferResult

data InferResult

Constructors

IR

Fields

ir_uniq :: Unique

ir_lvl :: TcLevel

ir_ref :: IORef (Maybe TcType)

type ExpSigmaType = ExpType

type ExpRhoType = ExpType

mkCheckExpType :: TcType -> ExpType

Make an ExpType suitable for checking.

data SyntaxOpType

What to expect for an argument to a rebindable-syntax operator. Quite like Type, but allows for holes to be filled in by tcSyntaxOp. The callback called from tcSyntaxOp gets a list of types; the meaning of these types is determined by a left-to-right depth-first traversal of the SyntaxOpType tree. So if you pass in

SynAny `SynFun` (SynList `SynFun` SynType Int) `SynFun` SynAny

you'll get three types back: one for the first SynAny, the element type of the list, and one for the last SynAny. You don't get anything for the SynType, because you've said positively that it should be an Int, and so it shall be.

You'll also get three multiplicities back: one for each function arrow. See also Note [Linear types] in Multiplicity.

This is defined here to avoid defining it in GHC.Tc.Gen.Expr boot file.

Constructors

SynAny
Any type

SynRho
A rho type, skolemised or instantiated as appropriate

SynList
A list type. You get back the element type of the list

SynFun SyntaxOpType SyntaxOpType infixr 0
A function.

SynType ExpType
A known type.

synKnownType :: TcType -> SyntaxOpType

Like SynType but accepts a regular TcType

mkSynFunTys :: [SyntaxOpType] -> ExpType -> SyntaxOpType

Like mkFunTys but for SyntaxOpType

newtype TcLevel

Constructors

TcLevel Int

topTcLevel :: TcLevel

pushTcLevel :: TcLevel -> TcLevel

isTopTcLevel :: TcLevel -> Bool

strictlyDeeperThan :: TcLevel -> TcLevel -> Bool

deeperThanOrSame :: TcLevel -> TcLevel -> Bool

sameDepthAs :: TcLevel -> TcLevel -> Bool

tcTypeLevel :: TcType -> TcLevel

tcTyVarLevel :: TcTyVar -> TcLevel

maxTcLevel :: TcLevel -> TcLevel -> TcLevel

promoteSkolem :: TcLevel -> TcTyVar -> TcTyVar

promoteSkolemX :: TcLevel -> TCvSubst -> TcTyVar -> (TCvSubst, TcTyVar)

Change the TcLevel in a skolem, extending a substitution

promoteSkolemsX :: TcLevel -> TCvSubst -> [TcTyVar] -> (TCvSubst, [TcTyVar])

data TcTyVarDetails

Constructors

SkolemTv TcLevel Bool

RuntimeUnk

MetaTv

Fields

mtv_info :: MetaInfo

mtv_ref :: IORef MetaDetails

mtv_tclvl :: TcLevel

pprTcTyVarDetails :: TcTyVarDetails -> SDoc

vanillaSkolemTv :: TcTyVarDetails

superSkolemTv :: TcTyVarDetails

data MetaDetails

Constructors

Flexi

Indirect TcType

data MetaInfo

Constructors

TauTv

TyVarTv

RuntimeUnkTv

CycleBreakerTv

isImmutableTyVar :: TyVar -> Bool

isSkolemTyVar :: TcTyVar -> Bool

isMetaTyVar :: TcTyVar -> Bool

isMetaTyVarTy :: TcType -> Bool

isTyVarTy :: Type -> Bool

tcIsTcTyVar :: TcTyVar -> Bool

isTyVarTyVar :: Var -> Bool

isOverlappableTyVar :: TcTyVar -> Bool

isTyConableTyVar :: TcTyVar -> Bool

isAmbiguousTyVar :: TcTyVar -> Bool

isCycleBreakerTyVar :: TcTyVar -> Bool

metaTyVarRef :: TyVar -> IORef MetaDetails

metaTyVarInfo :: TcTyVar -> MetaInfo

isFlexi :: MetaDetails -> Bool

isIndirect :: MetaDetails -> Bool

isRuntimeUnkSkol :: TyVar -> Bool

metaTyVarTcLevel :: TcTyVar -> TcLevel

setMetaTyVarTcLevel :: TcTyVar -> TcLevel -> TcTyVar

metaTyVarTcLevel_maybe :: TcTyVar -> Maybe TcLevel

isTouchableMetaTyVar :: TcLevel -> TcTyVar -> Bool

isPromotableMetaTyVar :: TcTyVar -> Bool

findDupTyVarTvs :: [(Name, TcTyVar)] -> [(Name, Name)]

mkTyVarNamePairs :: [TyVar] -> [(Name, TyVar)]

mkPhiTy :: [PredType] -> Type -> Type

mkInfSigmaTy :: [TyCoVar] -> [PredType] -> Type -> Type

Make a sigma ty where all type variables are Inferred. That is, they cannot be used with visible type application.

mkSpecSigmaTy :: [TyVar] -> [PredType] -> Type -> Type

Make a sigma ty where all type variables are "specified". That is, they can be used with visible type application

mkSigmaTy :: [TyCoVarBinder] -> [PredType] -> Type -> Type

mkTcAppTy :: Type -> Type -> Type

mkTcAppTys :: Type -> [Type] -> Type

mkTcCastTy :: Type -> Coercion -> Type

getTyVar :: String -> Type -> TyVar

Attempts to obtain the type variable underlying a Type, and panics with the given message if this is not a type variable type. See also getTyVar_maybe

tcSplitForAllTyVarBinder_maybe :: Type -> Maybe (TyVarBinder, Type)

tcSplitForAllTyVars :: Type -> ([TyVar], Type)

Like tcSplitPiTys, but splits off only named binders, returning just the tyvars.

tcSplitForAllInvisTyVars :: Type -> ([TyVar], Type)

Like tcSplitForAllTyVars, but only splits ForAllTys with Invisible type variable binders.

tcSplitSomeForAllTyVars :: (ArgFlag -> Bool) -> Type -> ([TyVar], Type)

Like tcSplitForAllTyVars, but only splits a ForAllTy if argf_pred argf is True, where argf is the visibility of the ForAllTy's binder and argf_pred is a predicate over visibilities provided as an argument to this function.

tcSplitForAllReqTVBinders :: Type -> ([TcReqTVBinder], Type)

Like tcSplitForAllTyVars, but only splits ForAllTys with Required type variable binders. All split tyvars are annotated with ().

tcSplitForAllInvisTVBinders :: Type -> ([TcInvisTVBinder], Type)

Like tcSplitForAllTyVars, but only splits ForAllTys with Invisible type variable binders. All split tyvars are annotated with their Specificity.

tcSplitPiTys :: Type -> ([TyBinder], Type)

Splits a forall type into a list of TyBinders and the inner type. Always succeeds, even if it returns an empty list.

tcSplitPiTy_maybe :: Type -> Maybe (TyBinder, Type)

Splits a type into a TyBinder and a body, if possible. Panics otherwise

tcSplitForAllTyVarBinders :: Type -> ([TyVarBinder], Type)

Like tcSplitForAllTyVars, but splits off only named binders.

tcSplitPhiTy :: Type -> (ThetaType, Type)

tcSplitPredFunTy_maybe :: Type -> Maybe (PredType, Type)

tcSplitFunTy_maybe :: Type -> Maybe (Scaled Type, Type)

tcSplitFunTys :: Type -> ([Scaled Type], Type)

tcFunArgTy :: Type -> Scaled Type

tcFunResultTy :: Type -> Type

tcFunResultTyN :: HasDebugCallStack => Arity -> Type -> Type

Strips off n *visible* arguments and returns the resulting type

tcSplitFunTysN :: Arity -> TcRhoType -> Either Arity ([Scaled TcSigmaType], TcSigmaType)

Split off exactly the specified number argument types Returns (Left m) if there are m missing arrows in the type (Right (tys,res)) if the type looks like t1 -> ... -> tn -> res

tcSplitTyConApp :: Type -> (TyCon, [Type])

tcSplitTyConApp_maybe :: HasCallStack => Type -> Maybe (TyCon, [Type])

Split a type constructor application into its type constructor and applied types. Note that this may fail in the case of a FunTy with an argument of unknown kind FunTy (e.g. FunTy (a :: k) Int. since the kind of a isn't of the form TYPE rep). Consequently, you may need to zonk your type before using this function.

This does *not* split types headed with (=>), as that's not a TyCon in the type-checker.

If you only need the TyCon, consider using tcTyConAppTyCon_maybe.

tcTyConAppTyCon :: Type -> TyCon

tcTyConAppTyCon_maybe :: Type -> Maybe TyCon

Like tcRepSplitTyConApp_maybe, but only returns the TyCon.

tcTyConAppArgs :: Type -> [Type]

tcSplitAppTy_maybe :: Type -> Maybe (Type, Type)

tcSplitAppTy :: Type -> (Type, Type)

tcSplitAppTys :: Type -> (Type, [Type])

tcRepSplitAppTy_maybe :: Type -> Maybe (Type, Type)

Does the AppTy split as in tcSplitAppTy_maybe, but assumes that any coreView stuff is already done. Refuses to look through (c => t)

tcRepGetNumAppTys :: Type -> Arity

Returns the number of arguments in the given type, without looking through synonyms. This is used only for error reporting. We don't look through synonyms because of #11313.

tcGetCastedTyVar_maybe :: Type -> Maybe (TyVar, CoercionN)

If the type is a tyvar, possibly under a cast, returns it, along with the coercion. Thus, the co is :: kind tv ~N kind type

tcGetTyVar_maybe :: Type -> Maybe TyVar

tcGetTyVar :: String -> Type -> TyVar

tcSplitSigmaTy :: Type -> ([TyVar], ThetaType, Type)

Split a sigma type into its parts. This only splits invisible type variable binders, as these are the only forms of binder that the typechecker will implicitly instantiate.

tcSplitNestedSigmaTys :: Type -> ([TyVar], ThetaType, Type)

Split a sigma type into its parts, going underneath as many arrows and foralls as possible. See Note [tcSplitNestedSigmaTys]

eqType :: Type -> Type -> Bool

Type equality on source types. Does not look through newtypes or PredTypes, but it does look through type synonyms. This first checks that the kinds of the types are equal and then checks whether the types are equal, ignoring casts and coercions. (The kind check is a recursive call, but since all kinds have type Type, there is no need to check the types of kinds.) See also Note [Non-trivial definitional equality] in GHC.Core.TyCo.Rep.

eqTypes :: [Type] -> [Type] -> Bool

Type equality on lists of types, looking through type synonyms but not newtypes.

nonDetCmpType :: Type -> Type -> Ordering

nonDetCmpTypes :: [Type] -> [Type] -> Ordering

eqTypeX :: RnEnv2 -> Type -> Type -> Bool

Compare types with respect to a (presumably) non-empty RnEnv2.

pickyEqType :: TcType -> TcType -> Bool

Like pickyEqTypeVis, but returns a Bool for convenience

tcEqType :: HasDebugCallStack => TcType -> TcType -> Bool

tcEqType implements typechecker equality, as described in Note [Typechecker equality vs definitional equality].

tcEqKind :: HasDebugCallStack => TcKind -> TcKind -> Bool

tcEqTypeNoKindCheck :: TcType -> TcType -> Bool

Just like tcEqType, but will return True for types of different kinds as long as their non-coercion structure is identical.

tcEqTypeVis :: TcType -> TcType -> Bool

Like tcEqType, but returns True if the visible part of the types are equal, even if they are really unequal (in the invisible bits)

tcEqTyConApps :: TyCon -> [Type] -> TyCon -> [Type] -> Bool

Check whether two TyConApps are the same; if the number of arguments are different, just checks the common prefix of arguments.

isSigmaTy :: TcType -> Bool

isRhoTy :: TcType -> Bool

isRhoExpTy :: ExpType -> Bool

Like isRhoTy, but also says True for Infer types

isOverloadedTy :: Type -> Bool

isFloatingTy :: Type -> Bool

Does a type represent a floating-point number?

isDoubleTy :: Type -> Bool

isFloatTy :: Type -> Bool

isIntTy :: Type -> Bool

isWordTy :: Type -> Bool

isStringTy :: Type -> Bool

Is a type String?

isIntegerTy :: Type -> Bool

isNaturalTy :: Type -> Bool

isBoolTy :: Type -> Bool

isUnitTy :: Type -> Bool

isCharTy :: Type -> Bool

isCallStackTy :: Type -> Bool

Is a type a CallStack?

isCallStackPred :: Class -> [Type] -> Maybe FastString

Is a PredType a CallStack implicit parameter?

If so, return the name of the parameter.

isTauTy :: Type -> Bool

isTauTyCon :: TyCon -> Bool

tcIsTyVarTy :: Type -> Bool

tcIsForAllTy :: Type -> Bool

Is this a ForAllTy with a named binder?

isPredTy :: HasDebugCallStack => Type -> Bool

isTyVarClassPred :: PredType -> Bool

checkValidClsArgs :: Bool -> Class -> [KindOrType] -> Bool

hasTyVarHead :: Type -> Bool

isRigidTy :: TcType -> Bool

deNoteType :: Type -> Type

orphNamesOfType :: Type -> NameSet

orphNamesOfCo :: Coercion -> NameSet

orphNamesOfTypes :: [Type] -> NameSet

orphNamesOfCoCon :: CoAxiom br -> NameSet

getDFunTyKey :: Type -> OccName

evVarPred :: EvVar -> PredType

mkMinimalBySCs :: forall a. (a -> PredType) -> [a] -> [a]

transSuperClasses :: PredType -> [PredType]

pickQuantifiablePreds :: TyVarSet -> TcThetaType -> TcThetaType

When inferring types, should we quantify over a given predicate? Generally true of classes; generally false of equality constraints. Equality constraints that mention quantified type variables and implicit variables complicate the story. See Notes [Inheriting implicit parameters] and [Quantifying over equality constraints]

pickCapturedPreds :: TyVarSet -> TcThetaType -> TcThetaType

immSuperClasses :: Class -> [Type] -> [PredType]

boxEqPred :: EqRel -> Type -> Type -> Maybe (Class, [Type])

isImprovementPred :: PredType -> Bool

Finding type instances

tcTyFamInsts :: Type -> [(TyCon, [Type])]

Finds outermost type-family applications occurring in a type, after expanding synonyms. In the list (F, tys) that is returned we guarantee that tys matches F's arity. For example, given type family F a :: * -> * (arity 1) calling tcTyFamInsts on (Maybe (F Int Bool) will return (F, [Int]), not (F, [Int,Bool])

This is important for its use in deciding termination of type instances (see #11581). E.g. type instance G [Int] = ...(F Int <big type>)... we don't need to take <big type> into account when asking if the calls on the RHS are smaller than the LHS

tcTyFamInstsAndVis :: Type -> [(Bool, TyCon, [Type])]

Like tcTyFamInsts, except that the output records whether the type family and its arguments occur as an invisible argument in some type application. This information is useful because it helps GHC know when to turn on -fprint-explicit-kinds during error reporting so that users can actually see the type family being mentioned.

As an example, consider:

class C a
data T (a :: k)
type family F a :: k
instance C (T @(F Int) (F Bool))

There are two occurrences of the type family F in that C instance, so tcTyFamInstsAndVis (C (T @(F Int) (F Bool))) will return:

[ (True,  F, [Int])
, (False, F, [Bool]) ]

F Int is paired with True since it appears as an invisible argument to C, whereas F Bool is paired with False since it appears an a visible argument to C.

See also Note [Kind arguments in error messages] in GHC.Tc.Errors.

tcTyConAppTyFamInstsAndVis :: TyCon -> [Type] -> [(Bool, TyCon, [Type])]

In an application of a TyCon to some arguments, find the outermost occurrences of type family applications within the arguments. This function will not consider the TyCon itself when checking for type family applications.

See tcTyFamInstsAndVis for more details on how this works (as this function is called inside of tcTyFamInstsAndVis).

isTyFamFree :: Type -> Bool

Check that a type does not contain any type family applications.

Finding "exact" (non-dead) type variables

exactTyCoVarsOfType :: Type -> TyCoVarSet

exactTyCoVarsOfTypes :: [Type] -> TyCoVarSet

anyRewritableTyVar :: Bool -> EqRel -> (EqRel -> TcTyVar -> Bool) -> TcType -> Bool

anyRewritableTyFamApp :: EqRel -> (EqRel -> TyCon -> [TcType] -> Bool) -> TcType -> Bool

anyRewritableCanEqLHS :: EqRel -> (EqRel -> TcTyVar -> Bool) -> (EqRel -> TyCon -> [TcType] -> Bool) -> TcType -> Bool

isFFIArgumentTy :: DynFlags -> Safety -> Type -> Validity

isFFIImportResultTy :: DynFlags -> Type -> Validity

isFFIExportResultTy :: Type -> Validity

isFFIExternalTy :: Type -> Validity

isFFIDynTy :: Type -> Type -> Validity

isFFIPrimArgumentTy :: DynFlags -> Type -> Validity

isFFIPrimResultTy :: DynFlags -> Type -> Validity

isFFILabelTy :: Type -> Validity

isFFITy :: Type -> Bool

isFunPtrTy :: Type -> Bool

tcSplitIOType_maybe :: Type -> Maybe (TyCon, Type)

type Kind = Type

The key type representing kinds in the compiler.

tcTypeKind :: HasDebugCallStack => Type -> Kind

liftedTypeKind :: Kind

constraintKind :: Kind

isLiftedTypeKind :: Kind -> Bool

This version considers Constraint to be the same as *. Returns True if the argument is equivalent to Type/Constraint and False otherwise. See Note [Kind Constraint and kind Type]

isUnliftedTypeKind :: Kind -> Bool

Returns True if the kind classifies unlifted types (like 'Int#') and False otherwise. Note that this returns False for levity-polymorphic kinds, which may be specialized to a kind that classifies unlifted types.

classifiesTypeWithValues :: Kind -> Bool

Does this classify a type allowed to have values? Responds True to things like *, #, TYPE Lifted, TYPE v, Constraint.

True of any sub-kind of OpenTypeKind

data Type

type PredType = Type

A type of the form p of constraint kind represents a value whose type is the Haskell predicate p, where a predicate is what occurs before the => in a Haskell type.

We use PredType as documentation to mark those types that we guarantee to have this kind.

It can be expanded into its representation, but:

The type checker must treat it as opaque

The rest of the compiler treats it as transparent

Consider these examples:

f :: (Eq a) => a -> Int
g :: (?x :: Int -> Int) => a -> Int
h :: (r\l) => {r} => {l::Int | r}

Here the Eq a and ?x :: Int -> Int and rl are all called "predicates"

type ThetaType = [PredType]

A collection of PredTypes

data TyCoBinder

A TyCoBinder represents an argument to a function. TyCoBinders can be dependent (Named) or nondependent (Anon). They may also be visible or not. See Note [TyCoBinders]

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, TyCoVarBinders, TyConBinders, and visibility] in GHC.Core.TyCo.Rep

Constructors

Invisible Specificity

Required

Bundled Patterns

pattern Specified :: ArgFlag

pattern Inferred :: ArgFlag

data AnonArgFlag

The non-dependent version of ArgFlag. See Note [AnonArgFlag] Appears here partly so that it's together with its friends ArgFlag and ForallVisFlag, but also because it is used in IfaceType, rather early in the compilation chain

Constructors

VisArg
Used for (->): an ordinary non-dependent arrow. The argument is visible in source code.

InvisArg
Used for (=>): a non-dependent predicate arrow. The argument is invisible in source code.

mkForAllTy :: TyCoVar -> ArgFlag -> Type -> Type

Like mkTyCoForAllTy, but does not check the occurrence of the binder See Note [Unused coercion variable in ForAllTy]

mkForAllTys :: [TyCoVarBinder] -> Type -> Type

Wraps foralls over the type using the provided TyCoVars from left to right

mkInvisForAllTys :: [InvisTVBinder] -> Type -> Type

Wraps foralls over the type using the provided InvisTVBinders from left to right

mkTyCoInvForAllTys :: [TyCoVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and Inferred, a common case

mkSpecForAllTys :: [TyVar] -> Type -> Type

Like mkForAllTys, but assumes all variables are dependent and Specified, a common case

mkTyCoInvForAllTy :: TyCoVar -> Type -> Type

Make a dependent forall over an Inferred variable

mkInfForAllTy :: TyVar -> Type -> Type

Like mkTyCoInvForAllTy, but tv should be a tyvar

mkInfForAllTys :: [TyVar] -> Type -> Type

Like mkTyCoInvForAllTys, but tvs should be a list of tyvar

mkVisFunTy :: Mult -> Type -> Type -> Type infixr 3

mkVisFunTys :: [Scaled Type] -> Type -> Type

Make nested arrow types

mkInvisFunTy :: Mult -> Type -> Type -> Type infixr 3

mkInvisFunTyMany :: Type -> Type -> Type infixr 3

mkVisFunTyMany :: Type -> Type -> Type infixr 3

Special, common, case: Arrow type with mult Many

mkVisFunTysMany :: [Type] -> Type -> Type

mkInvisFunTysMany :: [Type] -> Type -> Type

mkTyConApp :: TyCon -> [Type] -> Type

A key function: builds a TyConApp or FunTy as appropriate to its arguments. Applies its arguments to the constructor from left to right.

mkAppTy :: Type -> Type -> Type

Applies a type to another, as in e.g. k a

mkAppTys :: Type -> [Type] -> Type

mkTyConTy :: TyCon -> Type

Create the plain type constructor type which has been applied to no type arguments at all.

mkTyVarTy :: TyVar -> Type

mkTyVarTys :: [TyVar] -> [Type]

mkTyCoVarTy :: TyCoVar -> Type

mkTyCoVarTys :: [TyCoVar] -> [Type]

isClassPred :: PredType -> Bool

isEqPrimPred :: PredType -> Bool

isIPLikePred :: Type -> Bool

isEqPred :: PredType -> Bool

isEqPredClass :: Class -> Bool

mkClassPred :: Class -> [Type] -> PredType

tcSplitDFunTy :: Type -> ([TyVar], [Type], Class, [Type])

tcSplitDFunHead :: Type -> (Class, [Type])

tcSplitMethodTy :: Type -> ([TyVar], PredType, Type)

isRuntimeRepVar :: TyVar -> Bool

Is a tyvar of type RuntimeRep?

isKindLevPoly :: Kind -> Bool

Tests whether the given kind (which should look like TYPE x) is something other than a constructor tree (that is, constructors at every node). E.g. True of TYPE k, TYPE (F Int) False of TYPE 'LiftedRep

isVisibleBinder :: TyCoBinder -> Bool

Does this binder bind a visible argument?

isInvisibleBinder :: TyCoBinder -> Bool

Does this binder bind an invisible argument?

data TCvSubst

Type & coercion substitution

The following invariants must hold of a TCvSubst:

The in-scope set is needed only to guide the generation of fresh uniques

In particular, the kind of the type variables in the in-scope set is not relevant

The substitution is only applied ONCE! This is because in general such application will not reach a fixed point.

Constructors

TCvSubst InScopeSet TvSubstEnv CvSubstEnv

type TvSubstEnv = TyVarEnv Type

A substitution of Types for TyVars and Kinds for KindVars

emptyTCvSubst :: TCvSubst

mkEmptyTCvSubst :: InScopeSet -> TCvSubst

zipTvSubst :: HasDebugCallStack => [TyVar] -> [Type] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please!

mkTvSubstPrs :: [(TyVar, Type)] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please!

notElemTCvSubst :: Var -> TCvSubst -> Bool

unionTCvSubst :: TCvSubst -> TCvSubst -> TCvSubst

getTvSubstEnv :: TCvSubst -> TvSubstEnv

setTvSubstEnv :: TCvSubst -> TvSubstEnv -> TCvSubst

getTCvInScope :: TCvSubst -> InScopeSet

extendTCvInScope :: TCvSubst -> Var -> TCvSubst

extendTCvInScopeList :: TCvSubst -> [Var] -> TCvSubst

extendTCvInScopeSet :: TCvSubst -> VarSet -> TCvSubst

extendTvSubstAndInScope :: TCvSubst -> TyVar -> Type -> TCvSubst

lookupTyVar :: TCvSubst -> TyVar -> Maybe Type

extendTCvSubst :: TCvSubst -> TyCoVar -> Type -> TCvSubst

substTyVarBndr :: HasCallStack => TCvSubst -> TyVar -> (TCvSubst, TyVar)

extendTvSubst :: TCvSubst -> TyVar -> Type -> TCvSubst

isInScope :: Var -> TCvSubst -> Bool

mkTCvSubst :: InScopeSet -> (TvSubstEnv, CvSubstEnv) -> TCvSubst

mkTvSubst :: InScopeSet -> TvSubstEnv -> TCvSubst

Make a TCvSubst with specified tyvar subst and empty covar subst

zipTyEnv :: HasDebugCallStack => [TyVar] -> [Type] -> TvSubstEnv

zipCoEnv :: HasDebugCallStack => [CoVar] -> [Coercion] -> CvSubstEnv

substTy :: HasCallStack => TCvSubst -> Type -> Type

Substitute within a Type The substitution has to satisfy the invariants described in Note [The substitution invariant].

substTys :: HasCallStack => TCvSubst -> [Type] -> [Type]

Substitute within several Types The substitution has to satisfy the invariants described in Note [The substitution invariant].

substScaledTys :: HasCallStack => TCvSubst -> [Scaled Type] -> [Scaled Type]

substTyWith :: HasCallStack => [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst

substTyWithCoVars :: [CoVar] -> [Coercion] -> Type -> Type

Substitute covars within a type

substTyAddInScope :: TCvSubst -> Type -> Type

Substitute within a Type after adding the free variables of the type to the in-scope set. This is useful for the case when the free variables aren't already in the in-scope set or easily available. See also Note [The substitution invariant].

substTyUnchecked :: TCvSubst -> Type -> Type

Substitute within a Type disabling the sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substTysUnchecked :: TCvSubst -> [Type] -> [Type]

Substitute within several Types disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTysUnchecked to substTys and remove this function. Please don't use in new code.

substScaledTyUnchecked :: HasCallStack => TCvSubst -> Scaled Type -> Scaled Type

substThetaUnchecked :: TCvSubst -> ThetaType -> ThetaType

Substitute within a ThetaType disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substThetaUnchecked to substTheta and remove this function. Please don't use in new code.

substTyWithUnchecked :: [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substCoUnchecked :: TCvSubst -> Coercion -> Coercion

Substitute within a Coercion disabling sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substCoWithUnchecked :: [TyVar] -> [Type] -> Coercion -> Coercion

Coercion substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substTheta :: HasCallStack => TCvSubst -> ThetaType -> ThetaType

Substitute within a ThetaType The substitution has to satisfy the invariants described in Note [The substitution invariant].

isUnliftedType :: HasDebugCallStack => Type -> Bool

See Type for what an unlifted type is. Panics on levity polymorphic types; See mightBeUnliftedType for a more approximate predicate that behaves better in the presence of levity polymorphism.

isUnboxedTupleType :: Type -> Bool

isPrimitiveType :: Type -> Bool

Returns true of types that are opaque to Haskell.

tcView :: Type -> Maybe Type

Gives the typechecker view of a type. This unwraps synonyms but leaves Constraint alone. c.f. coreView, which turns Constraint into Type. Returns Nothing if no unwrapping happens. See also Note [coreView vs tcView]

coreView :: Type -> Maybe Type

This function strips off the top layer only of a type synonym application (if any) its underlying representation type. Returns Nothing if there is nothing to look through. This function considers Constraint to be a synonym of Type.

By being non-recursive and inlined, this case analysis gets efficiently joined onto the case analysis that the caller is already doing

tyCoVarsOfType :: Type -> TyCoVarSet

tyCoVarsOfTypes :: [Type] -> TyCoVarSet

closeOverKinds :: TyCoVarSet -> TyCoVarSet

tyCoFVsOfType :: Type -> FV

The worker for tyCoFVsOfType and tyCoFVsOfTypeList. The previous implementation used unionVarSet which is O(n+m) and can make the function quadratic. It's exported, so that it can be composed with other functions that compute free variables. See Note [FV naming conventions] in GHC.Utils.FV.

Eta-expanded because that makes it run faster (apparently) See Note [FV eta expansion] in GHC.Utils.FV for explanation.

tyCoFVsOfTypes :: [Type] -> FV

tyCoVarsOfTypeDSet :: Type -> DTyCoVarSet

tyCoFVsOfType that returns free variables of a type in a deterministic set. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in GHC.Utils.FV.

tyCoVarsOfTypesDSet :: [Type] -> DTyCoVarSet

Returns free variables of types, including kind variables as a deterministic set. For type synonyms it does not expand the synonym.

closeOverKindsDSet :: DTyVarSet -> DTyVarSet

Add the kind variables free in the kinds of the tyvars in the given set. Returns a deterministic set.

tyCoVarsOfTypeList :: Type -> [TyCoVar]

tyCoFVsOfType that returns free variables of a type in deterministic order. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in GHC.Utils.FV.

tyCoVarsOfTypesList :: [Type] -> [TyCoVar]

Returns free variables of types, including kind variables as a deterministically ordered list. For type synonyms it does not expand the synonym.

noFreeVarsOfType :: Type -> Bool

pprKind :: Kind -> SDoc

pprParendKind :: Kind -> SDoc

pprSigmaType :: Type -> SDoc

pprType :: Type -> SDoc

pprParendType :: Type -> SDoc

pprTypeApp :: TyCon -> [Type] -> SDoc

pprTheta :: ThetaType -> SDoc

pprParendTheta :: ThetaType -> SDoc

pprThetaArrowTy :: ThetaType -> SDoc

pprClassPred :: Class -> [Type] -> SDoc

pprTCvBndr :: TyCoVarBinder -> SDoc

pprTCvBndrs :: [TyCoVarBinder] -> SDoc

type TypeSize = IntWithInf

sizeType :: Type -> TypeSize

sizeTypes :: [Type] -> TypeSize

scopedSort :: [TyCoVar] -> [TyCoVar]

Do a topological sort on a list of tyvars, so that binders occur before occurrences E.g. given [ a::k, k::*, b::k ] it'll return a well-scoped list [ k::*, a::k, b::k ]

This is a deterministic sorting operation (that is, doesn't depend on Uniques).

It is also meant to be stable: that is, variables should not be reordered unnecessarily. This is specified in Note [ScopedSort] See also Note [Ordering of implicit variables] in GHC.Rename.HsType

tcTyConVisibilities :: TyCon -> [Bool]

For every arg a tycon can take, the returned list says True if the argument is taken visibly, and False otherwise. Ends with an infinite tail of Trues to allow for oversaturation.

isNextTyConArgVisible :: TyCon -> [Type] -> Bool

If the tycon is applied to the types, is the next argument visible?

isNextArgVisible :: TcType -> Bool

Should this type be applied to a visible argument?