GHC.Types.Basic

Documentation

data LeftOrRight

Constructors

CLeft

CRight

pickLR :: LeftOrRight -> (a, a) -> a

type ConTag = Int

A *one-index* constructor tag

Type of the tags associated with each constructor possibility or superclass selector

type ConTagZ = Int

A *zero-indexed* constructor tag

fIRST_TAG :: ConTag

Tags are allocated from here for real constructors or for superclass selectors

type Arity = Int

The number of value arguments that can be applied to a value before it does "real work". So: fib 100 has arity 0 x -> fib x has arity 1 See also Note [Definition of arity] in GHC.Core.Opt.Arity

type RepArity = Int

Representation Arity

The number of represented arguments that can be applied to a value before it does "real work". So: fib 100 has representation arity 0 x -> fib x has representation arity 1 (# x, y #) -> fib (x + y) has representation arity 2

type JoinArity = Int

The number of arguments that a join point takes. Unlike the arity of a function, this is a purely syntactic property and is fixed when the join point is created (or converted from a value). Both type and value arguments are counted.

type FullArgCount = Int

FullArgCount is the number of type or value arguments in an application, or the number of type or value binders in a lambda. Note: it includes both type and value arguments!

data Alignment

A power-of-two alignment

mkAlignment :: Int -> Alignment

alignmentOf :: Int -> Alignment

alignmentBytes :: Alignment -> Int

data PromotionFlag

Constructors

NotPromoted

IsPromoted

isPromoted :: PromotionFlag -> Bool

data FunctionOrData

Constructors

IsFunction

IsData

data RecFlag

Recursivity Flag

Constructors

Recursive

NonRecursive

isRec :: RecFlag -> Bool

isNonRec :: RecFlag -> Bool

boolToRecFlag :: Bool -> RecFlag

data Origin

Constructors

FromSource

Generated

isGenerated :: Origin -> Bool

type RuleName = FastString

pprRuleName :: RuleName -> SDoc

data TopLevelFlag

Constructors

TopLevel

NotTopLevel

isTopLevel :: TopLevelFlag -> Bool

isNotTopLevel :: TopLevelFlag -> Bool

data OverlapFlag

The semantics allowed for overlapping instances for a particular instance. See Note [Safe Haskell isSafeOverlap] in GHC.Core.InstEnv for a explanation of the isSafeOverlap field.

AnnKeywordId : AnnOpen '{-# OVERLAPPABLE' or '{-# OVERLAPPING' or '{-# OVERLAPS' or '{-# INCOHERENT', AnnClose `#-}`,

Constructors

OverlapFlag

Fields

overlapMode :: OverlapMode

isSafeOverlap :: Bool

data OverlapMode

Constructors

NoOverlap SourceText
This instance must not overlap another NoOverlap instance. However, it may be overlapped by Overlapping instances, and it may overlap Overlappable instances.

Overlappable SourceText
Silently ignore this instance if you find a more specific one that matches the constraint you are trying to resolve

Example: constraint (Foo [Int]) instance Foo [Int] instance {-# OVERLAPPABLE #-} Foo [a]

Since the second instance has the Overlappable flag, the first instance will be chosen (otherwise its ambiguous which to choose)

Overlapping SourceText
Silently ignore any more general instances that may be used to solve the constraint.

Example: constraint (Foo [Int]) instance {-# OVERLAPPING #-} Foo [Int] instance Foo [a]

Since the first instance has the Overlapping flag, the second---more general---instance will be ignored (otherwise it is ambiguous which to choose)

Overlaps SourceText
Equivalent to having both Overlapping and Overlappable flags.

Incoherent SourceText
Behave like Overlappable and Overlapping, and in addition pick an arbitrary one if there are multiple matching candidates, and don't worry about later instantiation

Example: constraint (Foo [b]) instance {-# INCOHERENT -} Foo [Int] instance Foo [a] Without the Incoherent flag, we'd complain that instantiating b would change which instance was chosen. See also Note [Incoherent instances] in GHC.Core.InstEnv

setOverlapModeMaybe :: OverlapFlag -> Maybe OverlapMode -> OverlapFlag

hasOverlappingFlag :: OverlapMode -> Bool

hasOverlappableFlag :: OverlapMode -> Bool

hasIncoherentFlag :: OverlapMode -> Bool

data Boxity

Constructors

Boxed

Unboxed

isBoxed :: Boxity -> Bool

data CbvMark

Should an argument be passed evaluated *and* tagged.

Constructors

MarkedCbv

NotMarkedCbv

isMarkedCbv :: CbvMark -> Bool

newtype PprPrec

A general-purpose pretty-printing precedence type.

Constructors

PprPrec Int

topPrec :: PprPrec

sigPrec :: PprPrec

opPrec :: PprPrec

funPrec :: PprPrec

starPrec :: PprPrec

appPrec :: PprPrec

maxPrec :: PprPrec

maybeParen :: PprPrec -> PprPrec -> SDoc -> SDoc

data TupleSort

Constructors

BoxedTuple

UnboxedTuple

ConstraintTuple

tupleSortBoxity :: TupleSort -> Boxity

boxityTupleSort :: Boxity -> TupleSort

tupleParens :: TupleSort -> SDoc -> SDoc

data UnboxedTupleOrSum

Are we dealing with an unboxed tuple or an unboxed sum?

Used when validity checking, see check_ubx_tuple_or_sum.

Constructors

UnboxedTupleType

UnboxedSumType

unboxedTupleOrSumExtension :: UnboxedTupleOrSum -> Extension

sumParens :: SDoc -> SDoc

pprAlternative

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> a
The things to be pretty printed

-> ConTag
Alternative (one-based)

-> Arity
Arity

-> SDoc
SDoc where the alternative havs been pretty printed and finally packed into a paragraph.

Pretty print an alternative in an unboxed sum e.g. "| a | |".

The OneShotInfo type

data OneShotInfo

If the Id is a lambda-bound variable then it may have lambda-bound variable info. Sometimes we know whether the lambda binding this variable is a "one-shot" lambda; that is, whether it is applied at most once.

This information may be useful in optimisation, as computations may safely be floated inside such a lambda without risk of duplicating work.

See also Note [OneShotInfo overview] above.

Constructors

NoOneShotInfo
No information

OneShotLam
The lambda is applied at most once.

noOneShotInfo :: OneShotInfo

It is always safe to assume that an Id has no lambda-bound variable information

hasNoOneShotInfo :: OneShotInfo -> Bool

isOneShotInfo :: OneShotInfo -> Bool

bestOneShot :: OneShotInfo -> OneShotInfo -> OneShotInfo

worstOneShot :: OneShotInfo -> OneShotInfo -> OneShotInfo

data OccInfo

identifier Occurrence Information

Constructors

ManyOccs
There are many occurrences, or unknown occurrences

Fields

occ_tail :: !TailCallInfo

IAmDead
Marks unused variables. Sometimes useful for lambda and case-bound variables.

OneOcc
Occurs exactly once (per branch), not inside a rule

Fields

occ_in_lam :: !InsideLam

occ_n_br :: !BranchCount

occ_int_cxt :: !InterestingCxt

occ_tail :: !TailCallInfo

IAmALoopBreaker
This identifier breaks a loop of mutually recursive functions. The field marks whether it is only a loop breaker due to a reference in a rule

Fields

occ_rules_only :: !RulesOnly

occ_tail :: !TailCallInfo

noOccInfo :: OccInfo

seqOccInfo :: OccInfo -> ()

zapFragileOcc :: OccInfo -> OccInfo

isOneOcc :: OccInfo -> Bool

isDeadOcc :: OccInfo -> Bool

isStrongLoopBreaker :: OccInfo -> Bool

isWeakLoopBreaker :: OccInfo -> Bool

isManyOccs :: OccInfo -> Bool

isNoOccInfo :: OccInfo -> Bool

strongLoopBreaker :: OccInfo

weakLoopBreaker :: OccInfo

data InsideLam

Inside Lambda

Constructors

IsInsideLam
Occurs inside a non-linear lambda Substituting a redex for this occurrence is dangerous because it might duplicate work.

NotInsideLam

type BranchCount = Int

oneBranch :: BranchCount

data InterestingCxt

Interesting Context

Constructors

IsInteresting
Function: is applied Data value: scrutinised by a case with at least one non-DEFAULT branch

NotInteresting

data TailCallInfo

Constructors

AlwaysTailCalled JoinArity

NoTailCallInfo

tailCallInfo :: OccInfo -> TailCallInfo

zapOccTailCallInfo :: OccInfo -> OccInfo

isAlwaysTailCalled :: OccInfo -> Bool

data EP a

Embedding Projection pair

Constructors

EP

Fields

fromEP :: a

toEP :: a

data DefMethSpec ty

Default Method Specification

Constructors

VanillaDM

GenericDM ty

data SwapFlag

Constructors

NotSwapped

IsSwapped

flipSwap :: SwapFlag -> SwapFlag

unSwap :: SwapFlag -> (a -> a -> b) -> a -> a -> b

isSwapped :: SwapFlag -> Bool

data CompilerPhase

Constructors

InitialPhase

Phase PhaseNum

FinalPhase

type PhaseNum = Int

Phase Number

beginPhase :: Activation -> CompilerPhase

nextPhase :: CompilerPhase -> CompilerPhase

laterPhase :: CompilerPhase -> CompilerPhase -> CompilerPhase

data Activation

Constructors

AlwaysActive

ActiveBefore SourceText PhaseNum

ActiveAfter SourceText PhaseNum

FinalActive

NeverActive

isActive :: CompilerPhase -> Activation -> Bool

competesWith :: Activation -> Activation -> Bool

isNeverActive :: Activation -> Bool

isAlwaysActive :: Activation -> Bool

activeInFinalPhase :: Activation -> Bool

activateAfterInitial :: Activation

activateDuringFinal :: Activation

activeAfter :: CompilerPhase -> Activation

data RuleMatchInfo

Rule Match Information

Constructors

ConLike

FunLike

isConLike :: RuleMatchInfo -> Bool

isFunLike :: RuleMatchInfo -> Bool

data InlineSpec

Inline Specification

Constructors

Inline SourceText

Inlinable SourceText

NoInline SourceText

Opaque SourceText

NoUserInlinePrag

noUserInlineSpec :: InlineSpec -> Bool

data InlinePragma

Constructors

InlinePragma

Fields

inl_src :: SourceText

inl_inline :: InlineSpec

inl_sat :: Maybe Arity

inl_act :: Activation

inl_rule :: RuleMatchInfo

defaultInlinePragma :: InlinePragma

alwaysInlinePragma :: InlinePragma

neverInlinePragma :: InlinePragma

dfunInlinePragma :: InlinePragma

isDefaultInlinePragma :: InlinePragma -> Bool

isInlinePragma :: InlinePragma -> Bool

isInlinablePragma :: InlinePragma -> Bool

isNoInlinePragma :: InlinePragma -> Bool

isOpaquePragma :: InlinePragma -> Bool

isAnyInlinePragma :: InlinePragma -> Bool

alwaysInlineConLikePragma :: InlinePragma

inlinePragmaSource :: InlinePragma -> SourceText

inlinePragmaName :: InlineSpec -> SDoc

Outputs string for pragma name for any of INLINEINLINABLENOINLINE. This differs from the Outputable instance for the InlineSpec type where the pragma name string as well as the accompanying SourceText (if any) is printed.

inlineSpecSource :: InlineSpec -> SourceText

inlinePragmaSpec :: InlinePragma -> InlineSpec

inlinePragmaSat :: InlinePragma -> Maybe Arity

inlinePragmaActivation :: InlinePragma -> Activation

inlinePragmaRuleMatchInfo :: InlinePragma -> RuleMatchInfo

setInlinePragmaActivation :: InlinePragma -> Activation -> InlinePragma

setInlinePragmaRuleMatchInfo :: InlinePragma -> RuleMatchInfo -> InlinePragma

pprInline :: InlinePragma -> SDoc

Pretty-print without displaying the user-specified InlineSpec.

pprInlineDebug :: InlinePragma -> SDoc

Pretty-print including the user-specified InlineSpec.

data UnfoldingSource

Constructors

VanillaSrc

StableUserSrc

StableSystemSrc

CompulsorySrc

isStableSource :: UnfoldingSource -> Bool

isStableUserSource :: UnfoldingSource -> Bool

isStableSystemSource :: UnfoldingSource -> Bool

isCompulsorySource :: UnfoldingSource -> Bool

data SuccessFlag

Constructors

Succeeded

Failed

succeeded :: SuccessFlag -> Bool

failed :: SuccessFlag -> Bool

successIf :: Bool -> SuccessFlag

data IntWithInf

An integer or infinity

infinity :: IntWithInf

A representation of infinity

treatZeroAsInf :: Int -> IntWithInf

Turn a positive number into an IntWithInf, where 0 represents infinity

subWithInf :: IntWithInf -> Int -> IntWithInf

Subtract an IntWithInf from an IntWithInf

mkIntWithInf :: Int -> IntWithInf

Inject any integer into an IntWithInf

intGtLimit :: Int -> IntWithInf -> Bool

data TypeOrKind

Flag to see whether we're type-checking terms or kind-checking types

Constructors

TypeLevel

KindLevel

isTypeLevel :: TypeOrKind -> Bool

isKindLevel :: TypeOrKind -> Bool

data Levity

Constructors

Lifted

Unlifted

mightBeLifted :: Maybe Levity -> Bool

mightBeUnlifted :: Maybe Levity -> Bool

data TypeOrConstraint

Constructors

TypeLike

ConstraintLike

data NonStandardDefaultingStrategy

Specify whether to default type variables of kind RuntimeRepLevityMultiplicity.

Constructors

DefaultNonStandardTyVars
Default type variables of the given kinds:

default RuntimeRep variables to LiftedRep

default Levity variables to Lifted

default Multiplicity variables to Many

TryNotToDefaultNonStandardTyVars
Try not to default type variables of the kinds RuntimeRepLevityMultiplicity.

Note that these might get defaulted anyway, if they are kind variables and `-XNoPolyKinds` is enabled.

data DefaultingStrategy

Specify whether to default kind variables, and type variables of kind RuntimeRepLevityMultiplicity.

Constructors

DefaultKindVars
Default kind variables:

default kind variables of kind Type to Type,

default RuntimeRepLevityMultiplicity kind variables to LiftedRepLiftedMany, respectively.

When this strategy is used, it means that we have determined that the variables we are considering defaulting are all kind variables.

Usually, we pass this option when -XNoPolyKinds is enabled.

NonStandardDefaulting NonStandardDefaultingStrategy
Default (or don't default) non-standard variables, of kinds RuntimeRep, Levity and Multiplicity.

defaultNonStandardTyVars :: DefaultingStrategy -> Bool

data ForeignSrcLang

Constructors

LangC

LangCxx

LangObjc

LangObjcxx

LangAsm

LangJs

RawObject

Orphan instances

Binary Boxity

Instance details

Methods

put_ :: BinHandle -> Boxity -> IO ()

put :: BinHandle -> Boxity -> IO (Bin Boxity)

get :: BinHandle -> IO Boxity

Binary PromotionFlag

Instance details

Methods

put_ :: BinHandle -> PromotionFlag -> IO ()

put :: BinHandle -> PromotionFlag -> IO (Bin PromotionFlag)

get :: BinHandle -> IO PromotionFlag

Outputable Boxity

Instance details

Methods

ppr :: Boxity -> SDoc

Outputable PromotionFlag

Instance details

Methods

ppr :: PromotionFlag -> SDoc