GHC.Types.Error

Messages

data Messages e

A collection of messages emitted by GHC during error reporting. A diagnostic message is typically a warning or an error. See Note [Messages].

type WarningMessages = Bag (MsgEnvelope DecoratedSDoc)

type ErrorMessages = Bag (MsgEnvelope DecoratedSDoc)

mkMessages :: Bag (MsgEnvelope e) -> Messages e

emptyMessages :: Messages e

isEmptyMessages :: Messages e -> Bool

addMessage :: MsgEnvelope e -> Messages e -> Messages e

unionMessages :: Messages e -> Messages e -> Messages e

Joins two collections of messages together.

data MsgEnvelope e

An envelope for GHC's facts about a running program, parameterised over the domain-specific (i.e. parsing, typecheck-renaming, etc) diagnostics.

To say things differently, GHC emits diagnostics about the running program, each of which is wrapped into a MsgEnvelope that carries specific information like where the error happened, its severity, etc. Finally, multiple MsgEnvelopes are aggregated into Messages that are returned to the user.

Constructors

MsgEnvelope

Fields

errMsgSpan :: SrcSpan

The SrcSpan is used for sorting errors into line-number order

errMsgContext :: PrintUnqualified

errMsgDiagnostic :: e

errMsgSeverity :: Severity

errMsgReason :: WarnReason

type WarnMsg = MsgEnvelope DecoratedSDoc

data SDoc

Represents a pretty-printable document.

To display an SDoc, use printSDoc, printSDocLn, bufLeftRenderSDoc, or renderWithContext. Avoid calling runSDoc directly as it breaks the abstraction layer.

data DecoratedSDoc

A DecoratedSDoc is isomorphic to a '[SDoc]' but it carries the invariant that the input '[SDoc]' needs to be rendered decorated into its final form, where the typical case would be adding bullets between each elements of the list. The type of decoration depends on the formatting function used, but in practice GHC uses the formatBulleted.

data Severity

Constructors

SevOutput

SevFatal

SevInteractive

SevDump
Log message intended for compiler developers No file/line/column stuff

SevInfo
Log messages intended for end users. No file/line/column stuff.

SevWarning

SevError
SevWarning and SevError are used for warnings and errors o The message has a file/line/column heading, plus "warning:" or "error:", added by mkLocMessags o Output is intended for end users

class RenderableDiagnostic a where

A class for types (typically errors and warnings) which can be "rendered" into an opaque DecoratedSDoc. For more information, see Note [Rendering Messages].

Methods

renderDiagnostic :: a -> DecoratedSDoc

pprMessageBag :: Bag SDoc -> SDoc

mkDecorated :: [SDoc] -> DecoratedSDoc

Creates a new DecoratedSDoc out of a list of SDoc.

mkLocMessage :: Severity -> SrcSpan -> SDoc -> SDoc

Make an unannotated error message with location info.

mkLocMessageAnn

Arguments

:: Maybe String
optional annotation

-> Severity
severity

-> SrcSpan
location

-> SDoc
message

-> SDoc

Make a possibly annotated error message with location info.

getSeverityColour :: Severity -> Scheme -> PprColour

getCaretDiagnostic :: Severity -> SrcSpan -> IO SDoc

makeIntoWarning :: WarnReason -> MsgEnvelope e -> MsgEnvelope e

Constructing individual errors

mkMsgEnvelope :: SrcSpan -> PrintUnqualified -> SDoc -> MsgEnvelope DecoratedSDoc

A short (one-line) error message

mkPlainMsgEnvelope :: SrcSpan -> SDoc -> MsgEnvelope DecoratedSDoc

Variant that doesn't care about qualified/unqualified names

mkErr :: SrcSpan -> PrintUnqualified -> e -> MsgEnvelope e

mkLongMsgEnvelope :: SrcSpan -> PrintUnqualified -> SDoc -> SDoc -> MsgEnvelope DecoratedSDoc

A long (multi-line) error message

mkWarnMsg :: SrcSpan -> PrintUnqualified -> SDoc -> MsgEnvelope DecoratedSDoc

A short (one-line) error message

mkPlainWarnMsg :: SrcSpan -> SDoc -> MsgEnvelope DecoratedSDoc

Variant that doesn't care about qualified/unqualified names

mkLongWarnMsg :: SrcSpan -> PrintUnqualified -> SDoc -> SDoc -> MsgEnvelope DecoratedSDoc

A long (multi-line) error message

Queries

isErrorMessage :: MsgEnvelope e -> Bool

isWarningMessage :: MsgEnvelope e -> Bool

getErrorMessages :: Messages e -> Bag (MsgEnvelope e)

getWarningMessages :: Messages e -> Bag (MsgEnvelope e)

partitionMessages :: Messages e -> (Bag (MsgEnvelope e), Bag (MsgEnvelope e))

Partitions the Messages and returns a tuple which first element are the warnings, and the second the errors.

errorsFound :: Messages e -> Bool