GHC.Types.FieldLabel

Documentation

type FieldLabelString = FastString

Field labels are just represented as strings; they are not necessarily unique (even within a module)

type FieldLabelEnv = DFastStringEnv FieldLabel

A map from labels to all the auxiliary information

data FieldLabel

Fields in an algebraic record type; see Note [FieldLabel].

Constructors

FieldLabel

Fields

flLabel :: FieldLabelString

User-visible label of the field

flHasDuplicateRecordFields :: DuplicateRecordFields

Was DuplicateRecordFields on in the defining module for this datatype?

flHasFieldSelector :: FieldSelectors

Was FieldSelectors enabled in the defining module for this datatype? See Note [NoFieldSelectors] in GHC.Rename.Env

flSelector :: Name

Record selector function

fieldSelectorOccName :: FieldLabelString -> OccName -> DuplicateRecordFields -> FieldSelectors -> OccName

Record selector OccNames are built from the underlying field name and the name of the first data constructor of the type, to support duplicate record field names. See Note [Why selector names include data constructors].

fieldLabelPrintableName :: FieldLabel -> Name

Undo the name mangling described in Note [FieldLabel] to produce a Name that has the user-visible OccName (but the selector's unique). This should be used only when generating output, when we want to show the label, but may need to qualify it with a module prefix.

data DuplicateRecordFields

Flag to indicate whether the DuplicateRecordFields extension is enabled.

Constructors

DuplicateRecordFields
Fields may be duplicated in a single module

NoDuplicateRecordFields
Fields must be unique within a module (the default)

data FieldSelectors

Flag to indicate whether the FieldSelectors extension is enabled.

Constructors

FieldSelectors
Selector functions are available (the default)

NoFieldSelectors
Selector functions are not available

flIsOverloaded :: FieldLabel -> Bool