GHC.Unit.Env

Documentation

data UnitEnv

Constructors

UnitEnv

Fields

ue_units :: !UnitState

Units

ue_home_unit :: !HomeUnit

Home unit

ue_platform :: !Platform

Platform

ue_namever :: !GhcNameVersion

GHC name/version (used for dynamic library suffix)

preloadUnitsInfo :: UnitEnv -> MaybeErr UnitErr [UnitInfo]

Lookup UnitInfo for every preload unit from the UnitState and for every unit used to instantiate the home unit.

preloadUnitsInfo' :: UnitEnv -> [UnitId] -> MaybeErr UnitErr [UnitInfo]

Lookup UnitInfo for every preload unit from the UnitState, for every unit used to instantiate the home unit, and for every unit explicitly passed in the given list of UnitId.