GHC.Unit.Module.Status

Documentation

data HscStatus

Status of a module compilation to machine code

Constructors

HscNotGeneratingCode ModIface ModDetails
Nothing to do.

HscUpToDate ModIface ModDetails
Nothing to do because code already exists.

HscUpdateBoot ModIface ModDetails
Update boot file result.

HscUpdateSig ModIface ModDetails
Generate signature file (backpack)

HscRecomp
Recompile this module.

Fields

hscs_guts :: CgGuts

Information for the code generator.

hscs_mod_location :: !ModLocation

Module info

hscs_partial_iface :: !PartialModIface

Partial interface

hscs_old_iface_hash :: !(Maybe Fingerprint)

Old interface hash for this compilation, if an old interface file exists. Pass to hscMaybeWriteIface when writing the interface to avoid updating the existing interface when the interface isn't changed.