GHC.Utils.Logger

Description

Logger

Documentation

data Logger

initLogger :: IO Logger

class HasLogger m where

Methods

getLogger :: m Logger

class ContainsLogger t where

Methods

extractLogger :: t -> Logger

type LogAction = DynFlags -> WarnReason -> Severity -> SrcSpan -> SDoc -> IO ()

type DumpAction = DynFlags -> PprStyle -> DumpFlag -> String -> DumpFormat -> SDoc -> IO ()

type TraceAction a = DynFlags -> String -> SDoc -> a -> a

data DumpFormat

Format of a dump

Dump formats are loosely defined: dumps may contain various additional headers and annotations and they may be partial. DumpFormat is mainly a hint (e.g. for syntax highlighters).

Constructors

FormatHaskell
Haskell

FormatCore
Core

FormatSTG
STG

FormatByteCode
ByteCode

FormatCMM
Cmm

FormatASM
Assembly code

FormatC
C code/header

FormatLLVM
LLVM bytecode

FormatText
Unstructured dump

putLogMsg :: Logger -> LogAction

Log something

putDumpMsg :: Logger -> DumpAction

Dump something

putTraceMsg :: Logger -> TraceAction a

Trace something

Hooks

popLogHook :: Logger -> Logger

Pop a log hook

pushLogHook :: (LogAction -> LogAction) -> Logger -> Logger

Push a log hook

popDumpHook :: Logger -> Logger

Pop a dump hook

pushDumpHook :: (DumpAction -> DumpAction) -> Logger -> Logger

Push a dump hook

popTraceHook :: Logger -> Logger

Pop a trace hook

pushTraceHook :: (forall a. TraceAction a -> TraceAction a) -> Logger -> Logger

Push a trace hook

makeThreadSafe :: Logger -> IO Logger

Make the logger thread-safe

Logging

jsonLogAction :: LogAction

defaultLogAction :: LogAction

defaultLogActionHPrintDoc :: DynFlags -> Bool -> Handle -> SDoc -> IO ()

Like defaultLogActionHPutStrDoc but appends an extra newline.

defaultLogActionHPutStrDoc :: DynFlags -> Bool -> Handle -> SDoc -> IO ()

The boolean arguments let's the pretty printer know if it can optimize indent by writing ascii ' ' characters without going through decoding.

Dumping

defaultDumpAction :: DumpCache -> LogAction -> DumpAction

Default action for dumpAction hook

withDumpFileHandle :: DumpCache -> DynFlags -> DumpFlag -> (Maybe Handle -> IO ()) -> IO ()

Run an action with the handle of a DumpFlag if we are outputting to a file, otherwise Nothing.

touchDumpFile :: Logger -> DynFlags -> DumpFlag -> IO ()

Ensure that a dump file is created even if it stays empty

dumpIfSet :: Logger -> DynFlags -> Bool -> String -> SDoc -> IO ()

dumpIfSet_dyn :: Logger -> DynFlags -> DumpFlag -> String -> DumpFormat -> SDoc -> IO ()

A wrapper around dumpAction. First check whether the dump flag is set Do nothing if it is unset

dumpIfSet_dyn_printer :: PrintUnqualified -> Logger -> DynFlags -> DumpFlag -> String -> DumpFormat -> SDoc -> IO ()

A wrapper around putDumpMsg. First check whether the dump flag is set Do nothing if it is unset

Unlike dumpIfSet_dyn, has a printer argument

Tracing

defaultTraceAction :: TraceAction a

Default action for traceAction hook