GHC.Utils.Outputable

Description

This module defines classes and functions for pretty-printing. It also exports a number of helpful debugging and other utilities such as trace and panic.

The interface to this module is very similar to the standard Hughes-PJ pretty printing module, except that it exports a number of additional functions that are rarely used, and works over the SDoc type.

Type classes

class Outputable a where

Class designating that some type has an SDoc representation

Methods

ppr :: a -> SDoc

class Outputable a => OutputableBndr a where

When we print a binder, we often want to print its type too. The OutputableBndr class encapsulates this idea.

Minimal complete definition

pprPrefixOcc, pprInfixOcc

Methods

pprBndr :: BindingSite -> a -> SDoc

pprPrefixOcc :: a -> SDoc

pprInfixOcc :: a -> SDoc

bndrIsJoin_maybe :: a -> Maybe Int

class OutputableP env a where

Outputable class with an additional environment value

See Note [The OutputableP class]

Methods

pdoc :: env -> a -> SDoc

class IsOutput doc where

A superclass for IsLine and IsDoc that provides an identity, empty, as well as access to the shared SDocContext.

See Note [The outputable class hierarchy] for more details.

Methods

empty :: doc

docWithContext :: (SDocContext -> doc) -> doc

class IsOutput doc => IsLine doc where

A class of types that represent a single logical line of text, with support for horizontal composition.

See Note [HLine versus HDoc] and Note [The outputable class hierarchy] for more details.

Minimal complete definition

char, text, ftext, ztext, (<>), (<+>), sep, fsep, dualLine

Methods

char :: Char -> doc

text :: String -> doc

ftext :: FastString -> doc

ztext :: FastZString -> doc

(<>) :: doc -> doc -> doc

Join two docs together horizontally without a gap.

(<+>) :: doc -> doc -> doc

Join two docs together horizontally with a gap between them.

sep :: [doc] -> doc

Separate: is either like hsep or like vcat, depending on what fits.

fsep :: [doc] -> doc

A paragraph-fill combinator. It's much like sep, only it keeps fitting things on one line until it can't fit any more.

hcat :: [doc] -> doc

Concatenate docs horizontally without gaps.

hsep :: [doc] -> doc

Concatenate docs horizontally with a space between each one.

dualLine :: SDoc -> HLine -> doc

Prints as either the given SDoc or the given HLine, depending on which type the result is instantiated to. This should generally be avoided; see Note [dualLine and dualDoc] for details.

class (IsOutput doc, IsLine (Line doc)) => IsDoc doc where

A class of types that represent a multiline document, with support for vertical composition.

See Note [HLine versus HDoc] and Note [The outputable class hierarchy] for more details.

Minimal complete definition

line, ($$), dualDoc

Associated Types

type Line doc = r | r -> doc

Methods

line :: Line doc -> doc

($$) :: doc -> doc -> doc

Join two docs together vertically. If there is no vertical overlap it "dovetails" the two onto one line.

lines_ :: [Line doc] -> doc

vcat :: [doc] -> doc

Concatenate docs vertically with dovetailing.

dualDoc :: SDoc -> HDoc -> doc

Prints as either the given SDoc or the given HDoc, depending on which type the result is instantiated to. This should generally be avoided; see Note [dualLine and dualDoc] for details.

data HLine

Represents a single line of output that can be efficiently printed directly to a Handle (actually a BufHandle). See Note [SDoc versus HDoc] and Note [HLine versus HDoc] for more details.

data HDoc

Represents a (possibly empty) sequence of lines that can be efficiently printed directly to a Handle (actually a BufHandle). See Note [SDoc versus HDoc] and Note [HLine versus HDoc] for more details.

Pretty printing combinators

data SDoc

Represents a pretty-printable document.

To display an SDoc, use printSDoc, printSDocLn, bufLeftRenderSDoc, or renderWithContext. Avoid calling runSDoc directly as it breaks the abstraction layer.

runSDoc :: SDoc -> SDocContext -> Doc

newtype PDoc a

Wrapper for types having a Outputable instance when an OutputableP instance is required.

Constructors

PDoc a

docToSDoc :: Doc -> SDoc

interppSP :: Outputable a => [a] -> SDoc

Returns the separated concatenation of the pretty printed things.

interpp'SP :: Outputable a => [a] -> SDoc

Returns the comma-separated concatenation of the pretty printed things.

interpp'SP' :: (a -> SDoc) -> [a] -> SDoc

pprQuotedList :: Outputable a => [a] -> SDoc

Returns the comma-separated concatenation of the quoted pretty printed things.

[x,y,z]  ==>  `x', `y', `z'

pprWithCommas

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> [a]
The things to be pretty printed

-> SDoc
SDoc where the things have been pretty printed, comma-separated and finally packed into a paragraph.

quotedListWithOr :: [SDoc] -> SDoc

quotedListWithNor :: [SDoc] -> SDoc

pprWithBars

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> [a]
The things to be pretty printed

-> SDoc
SDoc where the things have been pretty printed, bar-separated and finally packed into a paragraph.

spaceIfSingleQuote :: SDoc -> SDoc

isEmpty :: SDocContext -> SDoc -> Bool

nest :: Int -> SDoc -> SDoc

Indent SDoc some specified amount

ptext :: PtrString -> SDoc

int :: IsLine doc => Int -> doc

intWithCommas :: Integral a => a -> SDoc

integer :: IsLine doc => Integer -> doc

word :: Integer -> SDoc

float :: IsLine doc => Float -> doc

double :: IsLine doc => Double -> doc

rational :: Rational -> SDoc

doublePrec :: Int -> Double -> SDoc

doublePrec p n shows a floating point number n with p digits of precision after the decimal point.

parens :: IsLine doc => doc -> doc

cparen :: Bool -> SDoc -> SDoc

brackets :: IsLine doc => doc -> doc

braces :: IsLine doc => doc -> doc

quotes :: SDoc -> SDoc

quote :: SDoc -> SDoc

doubleQuotes :: IsLine doc => doc -> doc

angleBrackets :: IsLine doc => doc -> doc

semi :: IsLine doc => doc

comma :: IsLine doc => doc

colon :: IsLine doc => doc

dcolon :: SDoc

space :: IsLine doc => doc

equals :: IsLine doc => doc

dot :: IsLine doc => doc

vbar :: IsLine doc => doc

arrow :: SDoc

lollipop :: SDoc

larrow :: SDoc

darrow :: SDoc

arrowt :: SDoc

larrowt :: SDoc

arrowtt :: SDoc

larrowtt :: SDoc

lambda :: SDoc

lparen :: IsLine doc => doc

rparen :: IsLine doc => doc

lbrack :: IsLine doc => doc

rbrack :: IsLine doc => doc

lbrace :: IsLine doc => doc

rbrace :: IsLine doc => doc

underscore :: IsLine doc => doc

blankLine :: SDoc

forAllLit :: SDoc

bullet :: SDoc

($+$) :: SDoc -> SDoc -> SDoc

Join two SDoc together vertically

cat :: [SDoc] -> SDoc

A paragraph-fill combinator. It's much like sep, only it keeps fitting things on one line until it can't fit any more.

fcat :: [SDoc] -> SDoc

This behaves like fsep, but it uses <> for horizontal composition rather than <+>

hang

Arguments

:: SDoc
The header

-> Int
Amount to indent the hung body

-> SDoc
The hung body, indented and placed below the header

-> SDoc

hangNotEmpty :: SDoc -> Int -> SDoc -> SDoc

This behaves like hang, but does not indent the second document when the header is empty.

punctuate

Arguments

:: IsLine doc

=> doc
The punctuation

-> [doc]
The list that will have punctuation added between every adjacent pair of elements

-> [doc]
Punctuated list

ppWhen :: IsOutput doc => Bool -> doc -> doc

ppUnless :: IsOutput doc => Bool -> doc -> doc

ppWhenOption :: (SDocContext -> Bool) -> SDoc -> SDoc

ppUnlessOption :: IsLine doc => (SDocContext -> Bool) -> doc -> doc

speakNth :: Int -> SDoc

Converts an integer to a verbal index:

speakNth 1 = text "first"
speakNth 5 = text "fifth"
speakNth 21 = text "21st"

speakN :: Int -> SDoc

Converts an integer to a verbal multiplicity:

speakN 0 = text "none"
speakN 5 = text "five"
speakN 10 = text "10"

speakNOf :: Int -> SDoc -> SDoc

Converts an integer and object description to a statement about the multiplicity of those objects:

speakNOf 0 (text "melon") = text "no melons"
speakNOf 1 (text "melon") = text "one melon"
speakNOf 3 (text "melon") = text "three melons"

plural :: [a] -> SDoc

Determines the pluralisation suffix appropriate for the length of a list:

plural [] = char 's'
plural ["Hello"] = empty
plural ["Hello", "World"] = char 's'

singular :: [a] -> SDoc

Determines the singular verb suffix appropriate for the length of a list:

singular [] = empty
singular["Hello"] = char 's'
singular ["Hello", "World"] = empty

isOrAre :: [a] -> SDoc

Determines the form of to be appropriate for the length of a list:

isOrAre [] = text "are"
isOrAre ["Hello"] = text "is"
isOrAre ["Hello", "World"] = text "are"

doOrDoes :: [a] -> SDoc

Determines the form of to do appropriate for the length of a list:

doOrDoes [] = text "do"
doOrDoes ["Hello"] = text "does"
doOrDoes ["Hello", "World"] = text "do"

itsOrTheir :: [a] -> SDoc

Determines the form of possessive appropriate for the length of a list:

itsOrTheir [x]   = text "its"
itsOrTheir [x,y] = text "their"
itsOrTheir []    = text "their"  -- probably avoid this

thisOrThese :: [a] -> SDoc

Determines the form of subject appropriate for the length of a list:

thisOrThese [x]   = text "This"
thisOrThese [x,y] = text "These"
thisOrThese []    = text "These"  -- probably avoid this

hasOrHave :: [a] -> SDoc

"has" or "have" depending on the length of a list.

unicodeSyntax :: SDoc -> SDoc -> SDoc

coloured :: PprColour -> SDoc -> SDoc

Apply the given colour/style for the argument.

Only takes effect if colours are enabled.

keyword :: SDoc -> SDoc

Converting SDoc into strings and outputting it

printSDoc :: SDocContext -> Mode -> Handle -> SDoc -> IO ()

The analog of printDoc_ for SDoc, which tries to make sure the terminal doesn't get screwed up by the ANSI color codes if an exception is thrown during pretty-printing.

printSDocLn :: SDocContext -> Mode -> Handle -> SDoc -> IO ()

Like printSDoc but appends an extra newline.

bufLeftRenderSDoc :: SDocContext -> BufHandle -> SDoc -> IO ()

An efficient variant of printSDoc specialized for LeftMode that outputs to a BufHandle.

pprCode :: SDoc -> SDoc

showSDocOneLine :: SDocContext -> SDoc -> String

showSDocUnsafe :: SDoc -> String

showPprUnsafe :: Outputable a => a -> String

renderWithContext :: SDocContext -> SDoc -> String

pprDebugAndThen :: SDocContext -> (String -> a) -> SDoc -> SDoc -> a

pprInfixVar :: Bool -> SDoc -> SDoc

pprPrefixVar :: Bool -> SDoc -> SDoc

pprHsChar :: Char -> SDoc

Special combinator for showing character literals.

pprHsString :: FastString -> SDoc

Special combinator for showing string literals.

pprHsBytes :: ByteString -> SDoc

Special combinator for showing bytestring literals.

primFloatSuffix :: SDoc

primCharSuffix :: SDoc

primDoubleSuffix :: SDoc

primInt8Suffix :: SDoc

primWord8Suffix :: SDoc

primInt16Suffix :: SDoc

primWord16Suffix :: SDoc

primInt32Suffix :: SDoc

primWord32Suffix :: SDoc

primInt64Suffix :: SDoc

primWord64Suffix :: SDoc

primIntSuffix :: SDoc

primWordSuffix :: SDoc

pprPrimChar :: Char -> SDoc

Special combinator for showing unboxed literals.

pprPrimInt :: Integer -> SDoc

pprPrimWord :: Integer -> SDoc

pprPrimInt8 :: Integer -> SDoc

pprPrimWord8 :: Integer -> SDoc

pprPrimInt16 :: Integer -> SDoc

pprPrimWord16 :: Integer -> SDoc

pprPrimInt32 :: Integer -> SDoc

pprPrimWord32 :: Integer -> SDoc

pprPrimInt64 :: Integer -> SDoc

pprPrimWord64 :: Integer -> SDoc

pprFastFilePath :: FastString -> SDoc

pprFilePathString :: IsLine doc => FilePath -> doc

Normalise, escape and render a string representing a path

e.g. "c:\whatever"

pprModuleName :: IsLine doc => ModuleName -> doc

Controlling the style in which output is printed

data BindingSite

BindingSite is used to tell the thing that prints binder what language construct is binding the identifier. This can be used to decide how much info to print. Also see Note [Binding-site specific printing] in GHC.Core.Ppr

Constructors

LambdaBind
The x in (x. e)

CaseBind
The x in case scrut of x { (y,z) -> ... }

CasePatBind
The y,z in case scrut of x { (y,z) -> ... }

LetBind
The x in (let x = rhs in e)

data PprStyle

Constructors

PprUser NamePprCtx Depth Coloured

PprDump NamePprCtx

PprCode
Print code; either C or assembler

data NamePprCtx

When printing code that contains original names, we need to map the original names back to something the user understands. This is the purpose of the triple of functions that gets passed around when rendering SDoc.

Constructors

QueryQualify

Fields

queryQualifyName :: QueryQualifyName

queryQualifyModule :: QueryQualifyModule

queryQualifyPackage :: QueryQualifyPackage

queryPromotionTick :: QueryPromotionTick

type QueryQualifyName = Module -> OccName -> QualifyName

Given a Name's Module and OccName, decide whether and how to qualify it.

type QueryQualifyModule = Module -> Bool

For a given module, we need to know whether to print it with a package name to disambiguate it.

type QueryQualifyPackage = Unit -> Bool

For a given package, we need to know whether to print it with the component id to disambiguate it.

type QueryPromotionTick = PromotedItem -> Bool

Given a promoted data constructor, decide whether to print a tick to disambiguate the namespace.

data PromotedItem

Constructors

PromotedItemListSyntax IsEmptyOrSingleton

PromotedItemTupleSyntax

PromotedItemDataCon OccName

newtype IsEmptyOrSingleton

Constructors

IsEmptyOrSingleton Bool

isListEmptyOrSingleton :: [a] -> IsEmptyOrSingleton

data PromotionTickContext

Flags that affect whether a promotion tick is printed.

Constructors

PromTickCtx

Fields

ptcListTuplePuns :: !Bool

ptcPrintRedundantPromTicks :: !Bool

reallyAlwaysQualify :: NamePprCtx

reallyAlwaysQualifyNames :: QueryQualifyName

alwaysQualify :: NamePprCtx

alwaysQualifyNames :: QueryQualifyName

NB: This won't ever show package IDs

alwaysQualifyModules :: QueryQualifyModule

neverQualify :: NamePprCtx

neverQualifyNames :: QueryQualifyName

neverQualifyModules :: QueryQualifyModule

alwaysQualifyPackages :: QueryQualifyPackage

neverQualifyPackages :: QueryQualifyPackage

alwaysPrintPromTick :: QueryPromotionTick

data QualifyName

Constructors

NameUnqual

NameQual ModuleName

NameNotInScope1

NameNotInScope2

queryQual :: PprStyle -> NamePprCtx

sdocOption :: (SDocContext -> a) -> (a -> SDoc) -> SDoc

updSDocContext :: (SDocContext -> SDocContext) -> SDoc -> SDoc

data SDocContext

Constructors

SDC

Fields

sdocStyle :: !PprStyle

sdocColScheme :: !Scheme

sdocLastColour :: !PprColour

The most recently used colour. This allows nesting colours.

sdocShouldUseColor :: !Bool

sdocDefaultDepth :: !Int

sdocLineLength :: !Int

sdocCanUseUnicode :: !Bool

True if Unicode encoding is supported and not disabled by GHC_NO_UNICODE environment variable

sdocHexWordLiterals :: !Bool

sdocPprDebug :: !Bool

sdocPrintUnicodeSyntax :: !Bool

sdocPrintCaseAsLet :: !Bool

sdocPrintTypecheckerElaboration :: !Bool

sdocPrintAxiomIncomps :: !Bool

sdocPrintExplicitKinds :: !Bool

sdocPrintExplicitCoercions :: !Bool

sdocPrintExplicitRuntimeReps :: !Bool

sdocPrintExplicitForalls :: !Bool

sdocPrintPotentialInstances :: !Bool

sdocPrintEqualityRelations :: !Bool

sdocSuppressTicks :: !Bool

sdocSuppressTypeSignatures :: !Bool

sdocSuppressTypeApplications :: !Bool

sdocSuppressIdInfo :: !Bool

sdocSuppressCoercions :: !Bool

sdocSuppressCoercionTypes :: !Bool

sdocSuppressUnfoldings :: !Bool

sdocSuppressVarKinds :: !Bool

sdocSuppressUniques :: !Bool

sdocSuppressModulePrefixes :: !Bool

sdocSuppressStgExts :: !Bool

sdocSuppressStgReps :: !Bool

sdocErrorSpans :: !Bool

sdocStarIsType :: !Bool

sdocLinearTypes :: !Bool

sdocListTuplePuns :: !Bool

sdocPrintTypeAbbreviations :: !Bool

sdocUnitIdForUser :: !(FastString -> SDoc)

Used to map UnitIds to more friendly "package-version:component" strings while pretty-printing.

Use pprWithUnitState to set it. Users should never have to set it to pretty-print SDocs emitted by GHC, otherwise it's a bug. It's an internal field used to thread the UnitState so that the Outputable instance of UnitId can use it.

See Note [Pretty-printing UnitId] in GHC.Unit for more details.

Note that we use FastString instead of UnitId to avoid boring module inter-dependency issues.

sdocWithContext :: (SDocContext -> SDoc) -> SDoc

defaultSDocContext :: SDocContext

Default pretty-printing options

traceSDocContext :: SDocContext

getPprStyle :: (PprStyle -> SDoc) -> SDoc

withPprStyle :: PprStyle -> SDoc -> SDoc

setStyleColoured :: Bool -> PprStyle -> PprStyle

pprDeeper :: SDoc -> SDoc

pprDeeperList :: ([SDoc] -> SDoc) -> [SDoc] -> SDoc

Truncate a list that is longer than the current depth.

pprSetDepth :: Depth -> SDoc -> SDoc

codeStyle :: PprStyle -> Bool

userStyle :: PprStyle -> Bool

dumpStyle :: PprStyle -> Bool

qualName :: PprStyle -> QueryQualifyName

qualModule :: PprStyle -> QueryQualifyModule

qualPackage :: PprStyle -> QueryQualifyPackage

promTick :: PprStyle -> QueryPromotionTick

mkErrStyle :: NamePprCtx -> PprStyle

Style for printing error messages

defaultErrStyle :: PprStyle

Default style for error messages, when we don't know NamePprCtx It's a bit of a hack because it doesn't take into account what's in scope Only used for desugarer warnings, and typechecker errors in interface sigs

defaultDumpStyle :: PprStyle

mkDumpStyle :: NamePprCtx -> PprStyle

defaultUserStyle :: PprStyle

mkUserStyle :: NamePprCtx -> Depth -> PprStyle

cmdlineParserStyle :: PprStyle

data Depth

Constructors

AllTheWay

PartWay Int
0 => stop

DefaultDepth
Use sdocDefaultDepth field as depth

withUserStyle :: NamePprCtx -> Depth -> SDoc -> SDoc

withErrStyle :: NamePprCtx -> SDoc -> SDoc

ifPprDebug :: IsOutput doc => doc -> doc -> doc

Says what to do with and without -dppr-debug

whenPprDebug :: IsOutput doc => doc -> doc

Says what to do with -dppr-debug; without, return empty

getPprDebug :: IsOutput doc => (Bool -> doc) -> doc

Indicate if -dppr-debug mode is enabled

bPutHDoc :: BufHandle -> SDocContext -> HDoc -> IO ()