GHC.Utils.Panic

Description

Defines basic functions for printing error messages.

It's hard to put these functions anywhere else without causing some unnecessary loops in the module dependency graph.

GHC exception type

data GhcException

GHC's own exception type error messages all take the form:

     <location>: <error>
 

If the location is on the command line, or in GHC itself, then <location>="ghc". All of the error types below correspond to a <location> of "ghc", except for ProgramError (where the string is assumed to contain a location already, so we don't print one).

Constructors

Signal Int
Some other fatal signal (SIGHUP,SIGTERM)

UsageError String
Prints the short usage msg after the error

CmdLineError String
A problem with the command line arguments, but don't print usage.

Panic String
The impossible happened.

PprPanic String SDoc

Sorry String
The user tickled something that's known not to work yet, but we're not counting it as a bug.

PprSorry String SDoc

InstallationError String
An installation problem.

ProgramError String
An error in the user's code, probably.

PprProgramError String SDoc

showGhcException :: SDocContext -> GhcException -> ShowS

Append a description of the given exception to this string.

showGhcExceptionUnsafe :: GhcException -> ShowS

Append a description of the given exception to this string.

Note that this uses defaultSDocContext, which doesn't use the options set by the user via DynFlags.

throwGhcException :: GhcException -> a

throwGhcExceptionIO :: GhcException -> IO a

handleGhcException :: ExceptionMonad m => (GhcException -> m a) -> m a -> m a

Command error throwing patterns

pprPanic :: HasCallStack => String -> SDoc -> a

Throw an exception saying "bug in GHC" with a callstack

panicDoc :: String -> SDoc -> a

Throw an exception saying "bug in GHC"

sorryDoc :: String -> SDoc -> a

Throw an exception saying "this isn't finished yet"

pgmErrorDoc :: String -> SDoc -> a

Throw an exception saying "bug in pgm being compiled" (used for unusual program errors)

Assertions

assertPprPanic :: HasCallStack => SDoc -> a

Panic with an assertion failure, recording the given file and line number. Should typically be accessed with the ASSERT family of macros

assertPpr :: HasCallStack => Bool -> SDoc -> a -> a

assertPprMaybe :: HasCallStack => Maybe SDoc -> a -> a

assertPprM :: (HasCallStack, Monad m) => m Bool -> SDoc -> m ()

massertPpr :: (HasCallStack, Applicative m) => Bool -> SDoc -> m ()

Call stacks

callStackDoc :: HasCallStack => SDoc

prettyCallStackDoc :: CallStack -> SDoc

Exception utilities

class (Typeable e, Show e) => Exception e where

Minimal complete definition

Nothing

Methods

toException :: e -> SomeException

fromException :: SomeException -> Maybe e

displayException :: e -> String

backtraceDesired :: e -> Bool

showException :: Exception e => e -> String

Show an exception as a string.

safeShowException :: Exception e => e -> IO String

Show an exception which can possibly throw other exceptions. Used when displaying exception thrown within TH code.

try :: Exception e => IO a -> IO (Either e a)

tryMost :: IO a -> IO (Either SomeException a)

Like try, but pass through UserInterrupt and Panic exceptions. Used when we want soft failures when reading interface files, for example. TODO: I'm not entirely sure if this is catching what we really want to catch

throwTo :: Exception e => ThreadId -> e -> IO ()

withSignalHandlers :: ExceptionMonad m => m a -> m a

Temporarily install standard signal handlers for catching ^C, which just throw an exception in the current thread.

module GHC.Utils.Panic.Plain