GHC.Utils.TmpFs

Description

Temporary file-system management

Documentation

data TmpFs

Temporary file-system

initTmpFs :: IO TmpFs

Initialise an empty TmpFs

forkTmpFsFrom :: TmpFs -> IO TmpFs

Initialise an empty TmpFs sharing unique numbers and per-process temporary directories with the given TmpFs

mergeTmpFsInto :: TmpFs -> TmpFs -> IO ()

Merge the first TmpFs into the second.

The first TmpFs is returned emptied.

data FilesToClean

A collection of files that must be deleted before ghc exits.

Constructors

FilesToClean

Fields

ftcGhcSession :: !(Set FilePath)

Files that will be deleted at the end of runGhc(T)

ftcCurrentModule :: !(Set FilePath)

Files that will be deleted the next time cleanCurrentModuleTempFiles is called, or otherwise at the end of the session.

emptyFilesToClean :: FilesToClean

An empty FilesToClean

data TempFileLifetime

Used when a temp file is created. This determines which component Set of FilesToClean will get the temp file

Constructors

TFL_CurrentModule
A file with lifetime TFL_CurrentModule will be cleaned up at the end of upweep_mod

TFL_GhcSession
A file with lifetime TFL_GhcSession will be cleaned up at the end of runGhc(T)

cleanTempDirs :: Logger -> TmpFs -> DynFlags -> IO ()

cleanTempFiles :: Logger -> TmpFs -> DynFlags -> IO ()

Delete all files in tmp_files_to_clean.

cleanCurrentModuleTempFiles :: Logger -> TmpFs -> DynFlags -> IO ()

Delete all files in tmp_files_to_clean. That have lifetime TFL_CurrentModule. If a file must be cleaned eventually, but must survive a cleanCurrentModuleTempFiles, ensure it has lifetime TFL_GhcSession.

addFilesToClean :: TmpFs -> TempFileLifetime -> [FilePath] -> IO ()

Ensure that new_files are cleaned on the next call of cleanTempFiles or cleanCurrentModuleTempFiles, depending on lifetime. If any of new_files are already tracked, they will have their lifetime updated.

changeTempFilesLifetime :: TmpFs -> TempFileLifetime -> [FilePath] -> IO ()

Update the lifetime of files already being tracked. If any files are not being tracked they will be discarded.

newTempName :: Logger -> TmpFs -> DynFlags -> TempFileLifetime -> Suffix -> IO FilePath

newTempLibName :: Logger -> TmpFs -> DynFlags -> TempFileLifetime -> Suffix -> IO (FilePath, FilePath, String)

newTempDir :: Logger -> TmpFs -> DynFlags -> IO FilePath

withSystemTempDirectory

Arguments

:: String
Directory name template. See openTempFile.

-> (FilePath -> IO a)
Callback that can use the directory

-> IO a

Create and use a temporary directory in the system standard temporary directory.

Behaves exactly the same as withTempDirectory, except that the parent temporary directory will be that returned by getTemporaryDirectory.

withTempDirectory

Arguments

:: FilePath
Temp directory to create the directory in

-> String
Directory name template. See openTempFile.

-> (FilePath -> IO a)
Callback that can use the directory

-> IO a

Create and use a temporary directory.

Creates a new temporary directory inside the given directory, making use of the template. The temp directory is deleted after use. For example:

withTempDirectory "src" "sdist." $ \tmpDir -> do ...

The tmpDir will be a new subdirectory of the given directory, e.g. src/sdist.342.