HsBinds

Documentation

type HsLocalBinds id = HsLocalBindsLR id id

Haskell Local Bindings

type LHsLocalBinds id = Located (HsLocalBinds id)

Located Haskell local bindings

data HsLocalBindsLR idL idR

Haskell Local Bindings with separate Left and Right identifier types

Bindings in a 'let' expression or a 'where' clause

Constructors

HsValBinds (HsValBindsLR idL idR)
Haskell Value Bindings

HsIPBinds (HsIPBinds idR)
Haskell Implicit Parameter Bindings

EmptyLocalBinds
Empty Local Bindings

type LHsLocalBindsLR idL idR = Located (HsLocalBindsLR idL idR)

type HsValBinds id = HsValBindsLR id id

Haskell Value Bindings

data HsValBindsLR idL idR

Haskell Value bindings with separate Left and Right identifier types (not implicit parameters) Used for both top level and nested bindings May contain pattern synonym bindings

Constructors

ValBindsIn (LHsBindsLR idL idR) [LSig idR]
Value Bindings In

Before renaming RHS; idR is always RdrName Not dependency analysed Recursive by default

ValBindsOut [(RecFlag, LHsBinds idL)] [LSig Name]
Value Bindings Out

After renaming RHS; idR can be Name or Id Dependency analysed, later bindings in the list may depend on earlier ones.

type LHsBind id = LHsBindLR id id

Located Haskell Binding

type LHsBinds id = LHsBindsLR id id

Located Haskell Bindings

type HsBind id = HsBindLR id id

Haskell Binding

type LHsBindsLR idL idR = Bag (LHsBindLR idL idR)

Located Haskell Bindings with separate Left and Right identifier types

type LHsBindLR idL idR = Located (HsBindLR idL idR)

Located Haskell Binding with separate Left and Right identifier types

data HsBindLR idL idR

Haskell Binding with separate Left and Right id's

Constructors

FunBind
Function-like Binding

FunBind is used for both functions f x = e and variables f = x -> e and strict variables !x = x + 1

Reason 1: Special case for type inference: see tcMonoBinds.

Reason 2: Instance decls can only have FunBinds, which is convenient. If you change this, you'll need to change e.g. rnMethodBinds

But note that the form f :: a->a = ... parses as a pattern binding, just like (f :: a -> a) = ...

Strict bindings have their strictness recorded in the SrcStrictness of their MatchContext. See Note [Varieties of binding pattern matches] for details about the relationship between FunBind and PatBind.

AnnKeywordIds

AnnFunId, attached to each element of fun_matches

AnnEqual,AnnWhere, AnnOpen,AnnClose,

Fields

fun_id :: Located idL

fun_matches :: MatchGroup idR (LHsExpr idR)

The payload

fun_co_fn :: HsWrapper

Coercion from the type of the MatchGroup to the type of the Id. Example:

     f :: Int -> forall a. a -> a
     f x y = y

Then the MatchGroup will have type (Int -> a' -> a') (with a free type variable a'). The coercion will take a CoreExpr of this type and convert it to a CoreExpr of type Int -> forall a'. a' -> a' Notice that the coercion captures the free a'.

bind_fvs :: PostRn idL NameSet

After the renamer, this contains the locally-bound free variables of this defn. See Note [Bind free vars]

fun_tick :: [Tickish Id]

Ticks to put on the rhs, if any

PatBind
Pattern Binding

The pattern is never a simple variable; That case is done by FunBind. See Note [Varieties of binding pattern matches] for details about the relationship between FunBind and PatBind.

Fields

pat_lhs :: LPat idL

pat_rhs :: GRHSs idR (LHsExpr idR)

pat_rhs_ty :: PostTc idR Type

Type of the GRHSs

bind_fvs :: PostRn idL NameSet

After the renamer, this contains the locally-bound free variables of this defn. See Note [Bind free vars]

pat_ticks :: ([Tickish Id], [[Tickish Id]])

Ticks to put on the rhs, if any, and ticks to put on the bound variables.

VarBind
Variable Binding

Dictionary binding and suchlike. All VarBinds are introduced by the type checker

Fields

var_id :: idL

var_rhs :: LHsExpr idR

Located only for consistency

var_inline :: Bool

True = inline this binding regardless (used for implication constraints only)

AbsBinds
Abstraction Bindings

Fields

abs_tvs :: [TyVar]

abs_ev_vars :: [EvVar]

Includes equality constraints

abs_exports :: [ABExport idL]

AbsBinds only gets used when idL = idR after renaming, but these need to be idL's for the collect... code in HsUtil to have the right type

abs_ev_binds :: [TcEvBinds]

Evidence bindings Why a list? See TcInstDcls Note [Typechecking plan for instance declarations]

abs_binds :: LHsBinds idL

Typechecked user bindings

AbsBindsSig
Abstraction Bindings Signature

Fields

abs_tvs :: [TyVar]

abs_ev_vars :: [EvVar]

Includes equality constraints

abs_sig_export :: idL

abs_sig_prags :: TcSpecPrags

abs_sig_ev_bind :: TcEvBinds

abs_sig_bind :: LHsBind idL

PatSynBind (PatSynBind idL idR)

AnnKeywordId : AnnPattern, AnnLarrow,AnnEqual, AnnWhere AnnOpen '{',AnnClose '}'

data ABExport id

Abtraction Bindings Export

Constructors

ABE

Fields

abe_poly :: id

Any INLINE pragmas is attached to this Id

abe_mono :: id

abe_wrap :: HsWrapper

See Note [ABExport wrapper] Shape: (forall abs_tvs. abs_ev_vars => abe_mono) ~ abe_poly

abe_prags :: TcSpecPrags

SPECIALISE pragmas

data PatSynBind idL idR

AnnKeywordId : AnnPattern, AnnEqual,AnnLarrow AnnWhere,AnnOpen '{', AnnClose '}',

Pattern Synonym binding

Constructors

PSB

Fields

psb_id :: Located idL

Name of the pattern synonym

psb_fvs :: PostRn idR NameSet

See Note [Bind free vars]

psb_args :: HsPatSynDetails (Located idR)

Formal parameter names

psb_def :: LPat idR

Right-hand side

psb_dir :: HsPatSynDir idR

Directionality

pprLHsBinds :: (OutputableBndrId idL, OutputableBndrId idR) => LHsBindsLR idL idR -> SDoc

pprLHsBindsForUser :: (OutputableBndrId idL, OutputableBndrId idR, OutputableBndrId id2) => LHsBindsLR idL idR -> [LSig id2] -> [SDoc]

pprDeclList :: [SDoc] -> SDoc

emptyLocalBinds :: HsLocalBindsLR a b

isEmptyLocalBinds :: HsLocalBindsLR a b -> Bool

eqEmptyLocalBinds :: HsLocalBindsLR a b -> Bool

isEmptyValBinds :: HsValBindsLR a b -> Bool

emptyValBindsIn :: HsValBindsLR a b

emptyValBindsOut :: HsValBindsLR a b

emptyLHsBinds :: LHsBindsLR idL idR

isEmptyLHsBinds :: LHsBindsLR idL idR -> Bool

plusHsValBinds :: HsValBinds a -> HsValBinds a -> HsValBinds a

ppr_monobind :: (OutputableBndrId idL, OutputableBndrId idR) => HsBindLR idL idR -> SDoc

pprTicks :: SDoc -> SDoc -> SDoc

data HsIPBinds id

Haskell Implicit Parameter Bindings

Constructors

IPBinds [LIPBind id] TcEvBinds

isEmptyIPBinds :: HsIPBinds id -> Bool

type LIPBind id = Located (IPBind id)

Located Implicit Parameter Binding

May have AnnKeywordId : AnnSemi when in a list

data IPBind id

Implicit parameter bindings.

These bindings start off as (Left "x") in the parser and stay that way until after type-checking when they are replaced with (Right d), where "d" is the name of the dictionary holding the evidence for the implicit parameter.

AnnKeywordId : AnnEqual

Constructors

IPBind (Either (Located HsIPName) id) (LHsExpr id)

type LSig name = Located (Sig name)

Located Signature

data Sig name

Signatures and pragmas

Constructors

TypeSig [Located name] (LHsSigWcType name)
An ordinary type signature

f :: Num a => a -> a

After renaming, this list of Names contains the named and unnamed wildcards brought into scope by this signature. For a signature _ -> _a -> Bool, the renamer will give the unnamed wildcard _ a freshly generated name, e.g. _w. _w and the named wildcard _a are then both replaced with fresh meta vars in the type. Their names are stored in the type signature that brought them into scope, in this third field to be more specific.

AnnKeywordId : AnnDcolon, AnnComma

PatSynSig [Located name] (LHsSigType name)
A pattern synonym type signature

pattern Single :: () => (Show a) => a -> [a]

AnnKeywordId : AnnPattern, AnnDcolon,AnnForall AnnDot,AnnDarrow

ClassOpSig Bool [Located name] (LHsSigType name)
A signature for a class method False: ordinary class-method signature True: generic-default class method signature e.g. class C a where op :: a -> a -- Ordinary default op :: Eq a => a -> a -- Generic default No wildcards allowed here

AnnKeywordId : AnnDefault, AnnDcolon

IdSig Id
A type signature in generated code, notably the code generated for record selectors. We simply record the desired Id itself, replete with its name, type and IdDetails. Otherwise it's just like a type signature: there should be an accompanying binding

FixSig (FixitySig name)
An ordinary fixity declaration

    infixl 8 ***

AnnKeywordId : AnnInfix, AnnVal

InlineSig (Located name) InlinePragma
An inline pragma

{#- INLINE f #-}

AnnKeywordId : AnnOpen '{-# INLINE' and '[', AnnClose,AnnOpen, AnnVal,AnnTilde, AnnClose

SpecSig (Located name) [LHsSigType name] InlinePragma
A specialisation pragma

{-# SPECIALISE f :: Int -> Int #-}

AnnKeywordId : AnnOpen, AnnOpen '{-# SPECIALISE' and '[', AnnTilde, AnnVal, AnnClose ']' and '#-}', AnnDcolon

SpecInstSig SourceText (LHsSigType name)
A specialisation pragma for instance declarations only

{-# SPECIALISE instance Eq [Int] #-}

(Class tys); should be a specialisation of the current instance declaration

AnnKeywordId : AnnOpen, AnnInstance,AnnClose

MinimalSig SourceText (LBooleanFormula (Located name))
A minimal complete definition pragma

{-# MINIMAL a | (b, c | (d | e)) #-}

AnnKeywordId : AnnOpen, AnnVbar,AnnComma, AnnClose

SCCFunSig SourceText (Located name) (Maybe (Located StringLiteral))
A "set cost centre" pragma for declarations

{-# SCC funName #-}

or

{-# SCC funName "cost_centre_name" #-}

CompleteMatchSig SourceText (Located [Located name]) (Maybe (Located name))
A complete match pragma

{-# COMPLETE C, D [:: T] #-}

Used to inform the pattern match checker about additional complete matchings which, for example, arise from pattern synonym definitions.

type LFixitySig name = Located (FixitySig name)

Located Fixity Signature

data FixitySig name

Fixity Signature

Constructors

FixitySig [Located name] Fixity

data TcSpecPrags

Type checker Specialisation Pragmas

TcSpecPrags conveys SPECIALISE pragmas from the type checker to the desugarer

Constructors

IsDefaultMethod
Super-specialised: a default method should be macro-expanded at every call site

SpecPrags [LTcSpecPrag]

type LTcSpecPrag = Located TcSpecPrag

Located Type checker Specification Pragmas

data TcSpecPrag

Type checker Specification Pragma

Constructors

SpecPrag Id HsWrapper InlinePragma
The Id to be specialised, an wrapper that specialises the polymorphic function, and inlining spec for the specialised function

noSpecPrags :: TcSpecPrags

hasSpecPrags :: TcSpecPrags -> Bool

isDefaultMethod :: TcSpecPrags -> Bool

isFixityLSig :: LSig name -> Bool

isTypeLSig :: LSig name -> Bool

isSpecLSig :: LSig name -> Bool

isSpecInstLSig :: LSig name -> Bool

isPragLSig :: LSig name -> Bool

isInlineLSig :: LSig name -> Bool

isMinimalLSig :: LSig name -> Bool

isSCCFunSig :: LSig name -> Bool

isCompleteMatchSig :: LSig name -> Bool

hsSigDoc :: Sig name -> SDoc

ppr_sig :: OutputableBndrId name => Sig name -> SDoc

pragBrackets :: SDoc -> SDoc

pragSrcBrackets :: SourceText -> String -> SDoc -> SDoc

Using SourceText in case the pragma was spelled differently or used mixed case

pprVarSig :: OutputableBndr id => [id] -> SDoc -> SDoc

pprSpec :: OutputableBndr id => id -> SDoc -> InlinePragma -> SDoc

pprTcSpecPrags :: TcSpecPrags -> SDoc

pprMinimalSig :: OutputableBndr name => LBooleanFormula (Located name) -> SDoc

data HsPatSynDetails a

Haskell Pattern Synonym Details

Constructors

InfixPatSyn a a
Infix Pattern Synonym

PrefixPatSyn [a]
Prefix Pattern Synonym

RecordPatSyn [RecordPatSynField a]
Record Pattern Synonym

data RecordPatSynField a

Record Pattern Synonym Field

Constructors

RecordPatSynField

Fields

recordPatSynSelectorId :: a

recordPatSynPatVar :: a

data HsPatSynDir id

Haskell Pattern Synonym Direction

Constructors

Unidirectional

ImplicitBidirectional

ExplicitBidirectional (MatchGroup id (LHsExpr id))