HsDecls

Description

Abstract syntax of global declarations.

Definitions for: SynDecl and ConDecl, ClassDecl, InstDecl, DefaultDecl and ForeignDecl.

Toplevel declarations

data HsDecl id

A Haskell Declaration

Constructors

TyClD (TyClDecl id)
Type or Class Declaration

InstD (InstDecl id)
Instance declaration

DerivD (DerivDecl id)
Deriving declaration

ValD (HsBind id)
Value declaration

SigD (Sig id)
Signature declaration

DefD (DefaultDecl id)
'default' declaration

ForD (ForeignDecl id)
Foreign declaration

WarningD (WarnDecls id)
Warning declaration

AnnD (AnnDecl id)
Annotation declaration

RuleD (RuleDecls id)
Rule declaration

VectD (VectDecl id)
Vectorise declaration

SpliceD (SpliceDecl id)
Splice declaration (Includes quasi-quotes)

DocD DocDecl
Documentation comment declaration

RoleAnnotD (RoleAnnotDecl id)
Role annotation declaration

type LHsDecl id

Arguments

= Located (HsDecl id)
When in a list this may have

AnnKeywordId : AnnSemi

data HsDataDefn name

Haskell Data type Definition

Constructors

HsDataDefn
Declares a data type or newtype, giving its constructors data/newtype T a = constrs data/newtype instance T [a] = constrs

Fields

dd_ND :: NewOrData

dd_ctxt :: LHsContext name

Context

dd_cType :: Maybe (Located CType)

dd_kindSig :: Maybe (LHsKind name)

Optional kind signature.

(Just k) for a GADT-style data, or data instance decl, with explicit kind sig

Always Nothing for H98-syntax decls

dd_cons :: [LConDecl name]

Data constructors

For data T a = T1 | T2 a the LConDecls all have ConDeclH98. For data T a where { T1 :: T a } the LConDecls all have ConDeclGADT.

dd_derivs :: HsDeriving name

Optional 'deriving' claues

type HsDeriving name

Arguments

= Located [LHsDerivingClause name]
The optional deriving clauses of a data declaration. Clauses is plural because one can specify multiple deriving clauses using the -XDerivingStrategies language extension.

The list of LHsDerivingClauses corresponds to exactly what the user requested to derive, in order. If no deriving clauses were specified, the list is empty.

Haskell Deriving clause

data HsDerivingClause name

A single deriving clause of a data declaration.

AnnKeywordId : AnnDeriving, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

Constructors

HsDerivingClause

Fields

deriv_clause_strategy :: Maybe (Located DerivStrategy)

The user-specified strategy (if any) to use when deriving deriv_clause_tys.

deriv_clause_tys :: Located [LHsSigType name]

The types to derive.

It uses LHsSigTypes because, with -XGeneralizedNewtypeDeriving, we can mention type variables that aren't bound by the datatype, e.g.

data T b = ... deriving (C [a])

should produce a derived instance for C [a] (T b).

type LHsDerivingClause name = Located (HsDerivingClause name)

Class or type declarations

data TyClDecl name

A type or class declaration.

Constructors

FamDecl
type/data family T :: *->*

AnnKeywordId : AnnType, AnnData, AnnFamily,AnnDcolon, AnnWhere,AnnOpenP, AnnDcolon,AnnCloseP, AnnEqual,AnnRarrow, AnnVbar

Fields

tcdFam :: FamilyDecl name

SynDecl
type declaration

AnnKeywordId : AnnType, AnnEqual,

Fields

tcdLName :: Located name

Type constructor

tcdTyVars :: LHsQTyVars name

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdRhs :: LHsType name

RHS of type declaration

tcdFVs :: PostRn name NameSet

DataDecl
data declaration

AnnKeywordId : AnnData, AnnFamily, AnnNewType, AnnNewType,AnnDcolon AnnWhere,

Fields

tcdLName :: Located name

Type constructor

tcdTyVars :: LHsQTyVars name

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdDataDefn :: HsDataDefn name

tcdDataCusk :: PostRn name Bool

does this have a CUSK?

tcdFVs :: PostRn name NameSet

ClassDecl

AnnKeywordId : AnnClass, AnnWhere,AnnOpen, AnnClose

The tcdFDs will have AnnVbar, AnnComma AnnRarrow

Fields

tcdCtxt :: LHsContext name

Context...

tcdLName :: Located name

Type constructor

tcdTyVars :: LHsQTyVars name

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdFDs :: [Located (FunDep (Located name))]

Functional deps

tcdSigs :: [LSig name]

Methods' signatures

tcdMeths :: LHsBinds name

Default methods

tcdATs :: [LFamilyDecl name]

Associated types;

tcdATDefs :: [LTyFamDefltEqn name]

Associated type defaults

tcdDocs :: [LDocDecl]

Haddock docs

tcdFVs :: PostRn name NameSet

type LTyClDecl name = Located (TyClDecl name)

Located Declaration of a Type or Class

data TyClGroup name

Type or Class Group

Constructors

TyClGroup

Fields

group_tyclds :: [LTyClDecl name]

group_roles :: [LRoleAnnotDecl name]

group_instds :: [LInstDecl name]

mkTyClGroup :: [LTyClDecl name] -> [LInstDecl name] -> TyClGroup name

emptyTyClGroup :: TyClGroup name

tyClGroupTyClDecls :: [TyClGroup name] -> [LTyClDecl name]

tyClGroupInstDecls :: [TyClGroup name] -> [LInstDecl name]

tyClGroupRoleDecls :: [TyClGroup name] -> [LRoleAnnotDecl name]

isClassDecl :: TyClDecl name -> Bool

type class

isDataDecl :: TyClDecl name -> Bool

True = argument is a data/newtype declaration.

isSynDecl :: TyClDecl name -> Bool

type or type instance declaration

tcdName :: TyClDecl name -> name

isFamilyDecl :: TyClDecl name -> Bool

type/data family declaration

isTypeFamilyDecl :: TyClDecl name -> Bool

type family declaration

isDataFamilyDecl :: TyClDecl name -> Bool

data family declaration

isOpenTypeFamilyInfo :: FamilyInfo name -> Bool

open type family info

isClosedTypeFamilyInfo :: FamilyInfo name -> Bool

closed type family info

tyFamInstDeclName :: TyFamInstDecl name -> name

tyFamInstDeclLName :: TyFamInstDecl name -> Located name

countTyClDecls :: [TyClDecl name] -> (Int, Int, Int, Int, Int)

pprTyClDeclFlavour :: TyClDecl a -> SDoc

tyClDeclLName :: TyClDecl name -> Located name

tyClDeclTyVars :: TyClDecl name -> LHsQTyVars name

hsDeclHasCusk :: TyClDecl Name -> Bool

Does this declaration have a complete, user-supplied kind signature? See Note [Complete user-supplied kind signatures]

famDeclHasCusk

Arguments

:: Maybe Bool
if associated, does the enclosing class have a CUSK?

-> FamilyDecl name

-> Bool

Does this family declaration have a complete, user-supplied kind signature?

data FamilyDecl name

type Family Declaration

Constructors

FamilyDecl

AnnKeywordId : AnnType, AnnData, AnnFamily, AnnWhere, AnnOpenP, AnnDcolon, AnnCloseP, AnnEqual, AnnRarrow, AnnVbar

Fields

fdInfo :: FamilyInfo name

fdLName :: Located name

fdTyVars :: LHsQTyVars name

fdFixity :: LexicalFixity

fdResultSig :: LFamilyResultSig name

fdInjectivityAnn :: Maybe (LInjectivityAnn name)

type LFamilyDecl name = Located (FamilyDecl name)

Located type Family Declaration

Instance declarations

data InstDecl name

Instance Declaration

Constructors

ClsInstD

Fields

cid_inst :: ClsInstDecl name

DataFamInstD

Fields

dfid_inst :: DataFamInstDecl name

TyFamInstD

Fields

tfid_inst :: TyFamInstDecl name

type LInstDecl name = Located (InstDecl name)

Located Instance Declaration

data NewOrData

Constructors

NewType
newtype Blah ...

DataType
data Blah ...

data FamilyInfo name

Constructors

DataFamily

OpenTypeFamily

ClosedTypeFamily (Maybe [LTyFamInstEqn name])
Nothing if we're in an hs-boot file and the user said "type family Foo x where .."

data TyFamInstDecl name

Type Family Instance Declaration

Constructors

TyFamInstDecl

AnnKeywordId : AnnType, AnnInstance,

Fields

tfid_eqn :: LTyFamInstEqn name

tfid_fvs :: PostRn name NameSet

type LTyFamInstDecl name = Located (TyFamInstDecl name)

Located Type Family Instance Declaration

instDeclDataFamInsts :: [LInstDecl name] -> [DataFamInstDecl name]

data DataFamInstDecl name

Data Family Instance Declaration

Constructors

DataFamInstDecl

AnnKeywordId : AnnData, AnnNewType,AnnInstance, AnnDcolon AnnWhere,AnnOpen, AnnClose

Fields

dfid_tycon :: Located name

dfid_pats :: HsTyPats name

dfid_fixity :: LexicalFixity

Fixity used in the declaration

dfid_defn :: HsDataDefn name

dfid_fvs :: PostRn name NameSet

type LDataFamInstDecl name = Located (DataFamInstDecl name)

Located Data Family Instance Declaration

pprDataFamInstFlavour :: DataFamInstDecl name -> SDoc

data TyFamEqn name pats

Type Family Equation

One equation in a type family instance declaration See Note [Type family instance declarations in HsSyn]

Constructors

TyFamEqn

AnnKeywordId : AnnEqual

Fields

tfe_tycon :: Located name

tfe_pats :: pats

tfe_fixity :: LexicalFixity

Fixity used in the declaration

tfe_rhs :: LHsType name

type TyFamInstEqn name = TyFamEqn name (HsTyPats name)

Type Family Instance Equation

type LTyFamInstEqn name

Arguments

= Located (TyFamInstEqn name)
May have AnnKeywordId : AnnSemi when in a list

Located Type Family Instance Equation

type TyFamDefltEqn name = TyFamEqn name (LHsQTyVars name)

Type Family Default Equation

type LTyFamDefltEqn name = Located (TyFamDefltEqn name)

Located Type Family Default Equation

type HsTyPats name

Arguments

= HsImplicitBndrs name [LHsType name]
Type patterns (with kind and type bndrs) See Note [Family instance declaration binders]

Haskell Type Patterns

type LClsInstDecl name = Located (ClsInstDecl name)

Located Class Instance Declaration

data ClsInstDecl name

Class Instance Declaration

Constructors

ClsInstDecl

AnnKeywordId : AnnInstance, AnnWhere, AnnOpen,AnnClose,

Fields

cid_poly_ty :: LHsSigType name

cid_binds :: LHsBinds name

cid_sigs :: [LSig name]

cid_tyfam_insts :: [LTyFamInstDecl name]

cid_datafam_insts :: [LDataFamInstDecl name]

cid_overlap_mode :: Maybe (Located OverlapMode)

AnnKeywordId : AnnOpen, AnnClose,

Standalone deriving declarations

data DerivDecl name

Deriving Declaration

Constructors

DerivDecl

Fields

deriv_type :: LHsSigType name

deriv_strategy :: Maybe (Located DerivStrategy)

deriv_overlap_mode :: Maybe (Located OverlapMode)

AnnKeywordId : AnnDeriving, AnnInstance, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

type LDerivDecl name = Located (DerivDecl name)

Located Deriving Declaration

RULE declarations

type LRuleDecls name = Located (RuleDecls name)

Located Rule Declarations

data RuleDecls name

Rule Declarations

Constructors

HsRules

Fields

rds_src :: SourceText

rds_rules :: [LRuleDecl name]

data RuleDecl name

Rule Declaration

Constructors

HsRule (Located (SourceText, RuleName)) Activation [LRuleBndr name] (Located (HsExpr name)) (PostRn name NameSet) (Located (HsExpr name)) (PostRn name NameSet)

AnnKeywordId : AnnOpen,AnnTilde, AnnVal, AnnClose, AnnForall,AnnDot, AnnEqual,

type LRuleDecl name = Located (RuleDecl name)

Located Rule Declaration

data RuleBndr name

Rule Binder

Constructors

RuleBndr (Located name)

RuleBndrSig (Located name) (LHsSigWcType name)

AnnKeywordId : AnnOpen, AnnDcolon,AnnClose

type LRuleBndr name = Located (RuleBndr name)

Located Rule Binder

collectRuleBndrSigTys :: [RuleBndr name] -> [LHsSigWcType name]

flattenRuleDecls :: [LRuleDecls name] -> [LRuleDecl name]

pprFullRuleName :: Located (SourceText, RuleName) -> SDoc

VECTORISE declarations

data VectDecl name

Vectorise Declaration

Constructors

HsVect SourceText (Located name) (LHsExpr name)

AnnKeywordId : AnnOpen, AnnEqual,AnnClose

HsNoVect SourceText (Located name)

AnnKeywordId : AnnOpen, AnnClose

HsVectTypeIn SourceText Bool (Located name) (Maybe (Located name))

AnnKeywordId : AnnOpen, AnnType,AnnClose, AnnEqual

HsVectTypeOut Bool TyCon (Maybe TyCon)

HsVectClassIn SourceText (Located name)

AnnKeywordId : AnnOpen, AnnClass,AnnClose,

HsVectClassOut Class

HsVectInstIn (LHsSigType name)

HsVectInstOut ClsInst

type LVectDecl name = Located (VectDecl name)

Located Vectorise Declaration

lvectDeclName :: NamedThing name => LVectDecl name -> Name

lvectInstDecl :: LVectDecl name -> Bool

default declarations

data DefaultDecl name

Default Declaration

Constructors

DefaultDecl [LHsType name]

AnnKeywordIds : AnnDefault, AnnOpen,AnnClose

type LDefaultDecl name = Located (DefaultDecl name)

Located Default Declaration

Template haskell declaration splice

data SpliceExplicitFlag

Constructors

ExplicitSplice
= $(f x y)

ImplicitSplice
= f x y, i.e. a naked top level expression

data SpliceDecl id

Splice Declaration

Constructors

SpliceDecl (Located (HsSplice id)) SpliceExplicitFlag

type LSpliceDecl name = Located (SpliceDecl name)

Located Splice Declaration

Foreign function interface declarations

data ForeignDecl name

Foreign Declaration

Constructors

ForeignImport

Fields

fd_name :: Located name

fd_sig_ty :: LHsSigType name

fd_co :: PostTc name Coercion

fd_fi :: ForeignImport

ForeignExport

AnnKeywordId : AnnForeign, AnnImport,AnnExport, AnnDcolon

Fields

fd_name :: Located name

fd_sig_ty :: LHsSigType name

fd_co :: PostTc name Coercion

fd_fe :: ForeignExport

type LForeignDecl name = Located (ForeignDecl name)

Located Foreign Declaration

data ForeignImport

Constructors

CImport (Located CCallConv) (Located Safety) (Maybe Header) CImportSpec (Located SourceText)

data ForeignExport

Constructors

CExport (Located CExportSpec) (Located SourceText)

noForeignImportCoercionYet :: PlaceHolder

noForeignExportCoercionYet :: PlaceHolder

data CImportSpec

Constructors

CLabel CLabelString

CFunction CCallTarget

CWrapper

Data-constructor declarations

data ConDecl name

data T b = forall a. Eq a => MkT a b
  MkT :: forall b a. Eq a => MkT a b

data T b where
     MkT1 :: Int -> T Int

data T = Int MkT Int
       | MkT2

data T a where
     Int MkT Int :: T Int

AnnKeywordIds : AnnOpen, AnnDotdot,AnnCLose, AnnEqual,AnnVbar, AnnDarrow,AnnDarrow, AnnForall,AnnDot

data Constructor Declaration

Constructors

ConDeclGADT

Fields

con_names :: [Located name]

con_type :: LHsSigType name

The type after the ‘::’

con_doc :: Maybe LHsDocString

A possible Haddock comment.

ConDeclH98

Fields

con_name :: Located name

con_qvars :: Maybe (LHsQTyVars name)

con_cxt :: Maybe (LHsContext name)

User-written context (if any)

con_details :: HsConDeclDetails name

Arguments

con_doc :: Maybe LHsDocString

A possible Haddock comment.

type LConDecl name

Arguments

= Located (ConDecl name)
May have AnnKeywordId : AnnSemi when in a GADT constructor list

Located data Constructor Declaration

type HsConDeclDetails name = HsConDetails (LBangType name) (Located [LConDeclField name])

Haskell data Constructor Declaration Details

hsConDeclArgTys :: HsConDeclDetails name -> [LBangType name]

getConNames :: ConDecl name -> [Located name]

getConDetails :: ConDecl name -> HsConDeclDetails name

gadtDeclDetails :: LHsSigType name -> (HsConDeclDetails name, LHsType name, LHsContext name, [LHsTyVarBndr name])

Document comments

data DocDecl

Documentation comment Declaration

Constructors

DocCommentNext HsDocString

DocCommentPrev HsDocString

DocCommentNamed String HsDocString

DocGroup Int HsDocString

type LDocDecl = Located DocDecl

Located Documentation comment Declaration

docDeclDoc :: DocDecl -> HsDocString

Deprecations

data WarnDecl name

Warning pragma Declaration

Constructors

Warning [Located name] WarningTxt

type LWarnDecl name = Located (WarnDecl name)

Located Warning pragma Declaration

data WarnDecls name

Warning pragma Declarations

Constructors

Warnings

Fields

wd_src :: SourceText

wd_warnings :: [LWarnDecl name]

type LWarnDecls name = Located (WarnDecls name)

Located Warning Declarations

Annotations

data AnnDecl name

Annotation Declaration

Constructors

HsAnnotation SourceText (AnnProvenance name) (Located (HsExpr name))

AnnKeywordId : AnnOpen, AnnType AnnModule AnnClose

type LAnnDecl name = Located (AnnDecl name)

Located Annotation Declaration

data AnnProvenance name

Annotation Provenance

Constructors

ValueAnnProvenance (Located name)

TypeAnnProvenance (Located name)

ModuleAnnProvenance

annProvenanceName_maybe :: AnnProvenance name -> Maybe name

Role annotations

data RoleAnnotDecl name

Role Annotation Declaration

Constructors

RoleAnnotDecl (Located name) [Located (Maybe Role)]

AnnKeywordId : AnnType, AnnRole

type LRoleAnnotDecl name = Located (RoleAnnotDecl name)

Located Role Annotation Declaration

roleAnnotDeclName :: RoleAnnotDecl name -> name

Injective type families

data FamilyResultSig name

type Family Result Signature

Constructors

NoSig

AnnKeywordId :

KindSig (LHsKind name)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP

TyVarSig (LHsTyVarBndr name)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP, AnnEqual

type LFamilyResultSig name = Located (FamilyResultSig name)

Located type Family Result Signature

data InjectivityAnn name

If the user supplied an injectivity annotation it is represented using InjectivityAnn. At the moment this is a single injectivity condition - see Note [Injectivity annotation]. `Located name` stores the LHS of injectivity condition. `[Located name]` stores the RHS of injectivity condition. Example:

type family Foo a b c = r | r -> a c where ...

This will be represented as "InjectivityAnn r [a, c]"

Constructors

InjectivityAnn (Located name) [Located name]

AnnKeywordId : AnnRarrow, AnnVbar

type LInjectivityAnn name = Located (InjectivityAnn name)

Located Injectivity Annotation

resultVariableName :: FamilyResultSig a -> Maybe a

Maybe return name of the result type variable

Grouping

data HsGroup id

Haskell Group

A HsDecl is categorised into a HsGroup before being fed to the renamer.

Constructors

HsGroup

Fields

hs_valds :: HsValBinds id

hs_splcds :: [LSpliceDecl id]

hs_tyclds :: [TyClGroup id]

hs_derivds :: [LDerivDecl id]

hs_fixds :: [LFixitySig id]

hs_defds :: [LDefaultDecl id]

hs_fords :: [LForeignDecl id]

hs_warnds :: [LWarnDecls id]

hs_annds :: [LAnnDecl id]

hs_ruleds :: [LRuleDecls id]

hs_vects :: [LVectDecl id]

hs_docs :: [LDocDecl]

emptyRdrGroup :: HsGroup a

emptyRnGroup :: HsGroup a

appendGroups :: HsGroup a -> HsGroup a -> HsGroup a

hsGroupInstDecls :: HsGroup id -> [LInstDecl id]