HsDecls

Description

Abstract syntax of global declarations.

Definitions for: SynDecl and ConDecl, ClassDecl, InstDecl, DefaultDecl and ForeignDecl.

Toplevel declarations

data HsDecl id

A Haskell Declaration

Constructors

TyClD (TyClDecl id)
A type or class declaration.

InstD (InstDecl id)
An instance declaration.

DerivD (DerivDecl id)

ValD (HsBind id)

SigD (Sig id)

DefD (DefaultDecl id)

ForD (ForeignDecl id)

WarningD (WarnDecl id)

AnnD (AnnDecl id)

RuleD (RuleDecl id)

VectD (VectDecl id)

SpliceD (SpliceDecl id)

DocD DocDecl

QuasiQuoteD (HsQuasiQuote id)

RoleAnnotD (RoleAnnotDecl id)

type LHsDecl id = Located (HsDecl id)

data HsDataDefn name

Constructors

HsDataDefn
Declares a data type or newtype, giving its constructors data/newtype T a = constrs data/newtype instance T [a] = constrs

Fields

dd_ND :: NewOrData

dd_ctxt :: LHsContext name

Context

dd_cType :: Maybe CType

dd_kindSig :: Maybe (LHsKind name)

Optional kind signature.

(Just k) for a GADT-style data, or data instance decl, with explicit kind sig

Always Nothing for H98-syntax decls

dd_cons :: [LConDecl name]

Data constructors

For data T a = T1 | T2 a the LConDecls all have ResTyH98. For data T a where { T1 :: T a } the LConDecls all have ResTyGADT.

dd_derivs :: Maybe [LHsType name]

Derivings; Nothing => not specified, Just [] => derive exactly what is asked

These "types" must be of form forall ab. C ty1 ty2 Typically the foralls and ty args are empty, but they are non-empty for the newtype-deriving case

Class or type declarations

data TyClDecl name

A type or class declaration.

Constructors

ForeignType

Fields

tcdLName :: Located name

Type constructor

tcdExtName :: Maybe FastString

FamDecl
type/data family T :: *->*

Fields

tcdFam :: FamilyDecl name

SynDecl
type declaration

Fields

tcdLName :: Located name

Type constructor

tcdTyVars :: LHsTyVarBndrs name

Type variables; for an associated type these include outer binders

tcdRhs :: LHsType name

RHS of type declaration

tcdFVs :: NameSet

DataDecl
data declaration

Fields

tcdLName :: Located name

Type constructor

tcdTyVars :: LHsTyVarBndrs name

Type variables; for an associated type these include outer binders

tcdDataDefn :: HsDataDefn name

tcdFVs :: NameSet

ClassDecl

Fields

tcdCtxt :: LHsContext name

Context...

tcdLName :: Located name

Type constructor

tcdTyVars :: LHsTyVarBndrs name

Type variables; for an associated type these include outer binders

tcdFDs :: [Located (FunDep name)]

Functional deps

tcdSigs :: [LSig name]

Methods' signatures

tcdMeths :: LHsBinds name

Default methods

tcdATs :: [LFamilyDecl name]

Associated types; ie

tcdATDefs :: [LTyFamInstDecl name]

Associated type defaults

tcdDocs :: [LDocDecl]

Haddock docs

tcdFVs :: NameSet

type LTyClDecl name = Located (TyClDecl name)

data TyClGroup name

Constructors

TyClGroup

Fields

group_tyclds :: [LTyClDecl name]

group_roles :: [LRoleAnnotDecl name]

tyClGroupConcat :: [TyClGroup name] -> [LTyClDecl name]

mkTyClGroup :: [LTyClDecl name] -> TyClGroup name

isClassDecl :: TyClDecl name -> Bool

type class

isDataDecl :: TyClDecl name -> Bool

True = argument is a data/newtype declaration.

isSynDecl :: TyClDecl name -> Bool

type or type instance declaration

tcdName :: TyClDecl name -> name

isFamilyDecl :: TyClDecl name -> Bool

type/data family declaration

isTypeFamilyDecl :: TyClDecl name -> Bool

type family declaration

isDataFamilyDecl :: TyClDecl name -> Bool

data family declaration

isOpenTypeFamilyInfo :: FamilyInfo name -> Bool

open type family info

isClosedTypeFamilyInfo :: FamilyInfo name -> Bool

closed type family info

tyFamInstDeclName :: OutputableBndr name => TyFamInstDecl name -> name

tyFamInstDeclLName :: OutputableBndr name => TyFamInstDecl name -> Located name

countTyClDecls :: [TyClDecl name] -> (Int, Int, Int, Int, Int)

pprTyClDeclFlavour :: TyClDecl a -> SDoc

tyClDeclLName :: TyClDecl name -> Located name

tyClDeclTyVars :: OutputableBndr name => TyClDecl name -> LHsTyVarBndrs name

data FamilyDecl name

Constructors

FamilyDecl

Fields

fdInfo :: FamilyInfo name

fdLName :: Located name

fdTyVars :: LHsTyVarBndrs name

fdKindSig :: Maybe (LHsKind name)

type LFamilyDecl name = Located (FamilyDecl name)

Instance declarations

data InstDecl name

Constructors

ClsInstD

Fields

cid_inst :: ClsInstDecl name

DataFamInstD

Fields

dfid_inst :: DataFamInstDecl name

TyFamInstD

Fields

tfid_inst :: TyFamInstDecl name

type LInstDecl name = Located (InstDecl name)

data NewOrData

Constructors

NewType
newtype Blah ...

DataType
data Blah ...

data FamilyInfo name

Constructors

DataFamily

OpenTypeFamily

ClosedTypeFamily [LTyFamInstEqn name]

data TyFamInstDecl name

Constructors

TyFamInstDecl

Fields

tfid_eqn :: LTyFamInstEqn name

tfid_fvs :: NameSet

type LTyFamInstDecl name = Located (TyFamInstDecl name)

instDeclDataFamInsts :: [LInstDecl name] -> [DataFamInstDecl name]

data DataFamInstDecl name

Constructors

DataFamInstDecl

Fields

dfid_tycon :: Located name

dfid_pats :: HsWithBndrs [LHsType name]

Type patterns (with kind and type bndrs) See Note [Family instance declaration binders]

dfid_defn :: HsDataDefn name

dfid_fvs :: NameSet

type LDataFamInstDecl name = Located (DataFamInstDecl name)

pprDataFamInstFlavour :: DataFamInstDecl name -> SDoc

data TyFamInstEqn name

One equation in a type family instance declaration

Constructors

TyFamInstEqn

Fields

tfie_tycon :: Located name

tfie_pats :: HsWithBndrs [LHsType name]

Type patterns (with kind and type bndrs) See Note [Family instance declaration binders]

tfie_rhs :: LHsType name

type LTyFamInstEqn name = Located (TyFamInstEqn name)

type LClsInstDecl name = Located (ClsInstDecl name)

data ClsInstDecl name

Constructors

ClsInstDecl

Fields

cid_poly_ty :: LHsType name

cid_binds :: LHsBinds name

cid_sigs :: [LSig name]

cid_tyfam_insts :: [LTyFamInstDecl name]

cid_datafam_insts :: [LDataFamInstDecl name]

Standalone deriving declarations

data DerivDecl name

Constructors

DerivDecl

Fields

deriv_type :: LHsType name

type LDerivDecl name = Located (DerivDecl name)

RULE declarations

data RuleDecl name

Constructors

HsRule RuleName Activation [RuleBndr name] (Located (HsExpr name)) NameSet (Located (HsExpr name)) NameSet

type LRuleDecl name = Located (RuleDecl name)

data RuleBndr name

Constructors

RuleBndr (Located name)

RuleBndrSig (Located name) (HsWithBndrs (LHsType name))

collectRuleBndrSigTys :: [RuleBndr name] -> [HsWithBndrs (LHsType name)]

VECTORISE declarations

data VectDecl name

Constructors

HsVect (Located name) (LHsExpr name)

HsNoVect (Located name)

HsVectTypeIn Bool (Located name) (Maybe (Located name))

HsVectTypeOut Bool TyCon (Maybe TyCon)

HsVectClassIn (Located name)

HsVectClassOut Class

HsVectInstIn (LHsType name)

HsVectInstOut ClsInst

type LVectDecl name = Located (VectDecl name)

lvectDeclName :: NamedThing name => LVectDecl name -> Name

lvectInstDecl :: LVectDecl name -> Bool

default declarations

data DefaultDecl name

Constructors

DefaultDecl [LHsType name]

type LDefaultDecl name = Located (DefaultDecl name)

Template haskell declaration splice

data SpliceDecl id

Constructors

SpliceDecl (Located (HsSplice id)) HsExplicitFlag

type LSpliceDecl name = Located (SpliceDecl name)

Foreign function interface declarations

data ForeignDecl name

Constructors

ForeignImport (Located name) (LHsType name) Coercion ForeignImport

ForeignExport (Located name) (LHsType name) Coercion ForeignExport

type LForeignDecl name = Located (ForeignDecl name)

data ForeignImport

Constructors

CImport CCallConv Safety (Maybe Header) CImportSpec

data ForeignExport

Constructors

CExport CExportSpec

noForeignImportCoercionYet :: Coercion

noForeignExportCoercionYet :: Coercion

data CImportSpec

Constructors

CLabel CLabelString

CFunction CCallTarget

CWrapper

Data-constructor declarations

data ConDecl name

Constructors

ConDecl

Fields

con_name :: Located name

Constructor name. This is used for the DataCon itself, and for the user-callable wrapper Id.

con_explicit :: HsExplicitFlag

Is there an user-written forall? (cf. HsForAllTy)

con_qvars :: LHsTyVarBndrs name

Type variables. Depending on con_res this describes the following entities

ResTyH98: the constructor's *existential* type variables

ResTyGADT: *all* the constructor's quantified type variables

If con_explicit is Implicit, then con_qvars is irrelevant until after renaming.

con_cxt :: LHsContext name

The context. This does not include the "stupid theta" which lives only in the TyData decl.

con_details :: HsConDeclDetails name

The main payload

con_res :: ResType (LHsType name)

Result type of the constructor

con_doc :: Maybe LHsDocString

A possible Haddock comment.

con_old_rec :: Bool

TEMPORARY field; True = user has employed now-deprecated syntax for GADT-style record decl C { blah } :: T a b Remove this when we no longer parse this stuff, and hence do not need to report decprecated use

type LConDecl name = Located (ConDecl name)

data ResType ty

Constructors

ResTyH98

ResTyGADT ty

type HsConDeclDetails name = HsConDetails (LBangType name) [ConDeclField name]

hsConDeclArgTys :: HsConDeclDetails name -> [LBangType name]

Document comments

data DocDecl

Constructors

DocCommentNext HsDocString

DocCommentPrev HsDocString

DocCommentNamed String HsDocString

DocGroup Int HsDocString

type LDocDecl = Located DocDecl

docDeclDoc :: DocDecl -> HsDocString

Deprecations

data WarnDecl name

Constructors

Warning name WarningTxt

type LWarnDecl name = Located (WarnDecl name)

Annotations

data AnnDecl name

Constructors

HsAnnotation (AnnProvenance name) (Located (HsExpr name))

type LAnnDecl name = Located (AnnDecl name)

data AnnProvenance name

Constructors

ValueAnnProvenance name

TypeAnnProvenance name

ModuleAnnProvenance

annProvenanceName_maybe :: AnnProvenance name -> Maybe name

Role annotations

data RoleAnnotDecl name

Constructors

RoleAnnotDecl (Located name) [Located (Maybe Role)]

type LRoleAnnotDecl name = Located (RoleAnnotDecl name)

roleAnnotDeclName :: RoleAnnotDecl name -> name

Grouping

data HsGroup id

A HsDecl is categorised into a HsGroup before being fed to the renamer.

Constructors

HsGroup

Fields

hs_valds :: HsValBinds id

hs_splcds :: [LSpliceDecl id]

hs_tyclds :: [TyClGroup id]

hs_instds :: [LInstDecl id]

hs_derivds :: [LDerivDecl id]

hs_fixds :: [LFixitySig id]

hs_defds :: [LDefaultDecl id]

hs_fords :: [LForeignDecl id]

hs_warnds :: [LWarnDecl id]

hs_annds :: [LAnnDecl id]

hs_ruleds :: [LRuleDecl id]

hs_vects :: [LVectDecl id]

hs_docs :: [LDocDecl]

emptyRdrGroup :: HsGroup a

emptyRnGroup :: HsGroup a

appendGroups :: HsGroup a -> HsGroup a -> HsGroup a