HsDecls

Description

Abstract syntax of global declarations.

Definitions for: SynDecl and ConDecl, ClassDecl, InstDecl, DefaultDecl and ForeignDecl.

Toplevel declarations

data HsDecl id

A Haskell Declaration

Constructors

TyClD (TyClDecl id)
Type or Class Declaration

InstD (InstDecl id)
Instance declaration

DerivD (DerivDecl id)
Deriving declaration

ValD (HsBind id)
Value declaration

SigD (Sig id)
Signature declaration

DefD (DefaultDecl id)
'default' declaration

ForD (ForeignDecl id)
Foreign declaration

WarningD (WarnDecls id)
Warning declaration

AnnD (AnnDecl id)
Annotation declaration

RuleD (RuleDecls id)
Rule declaration

VectD (VectDecl id)
Vectorise declaration

SpliceD (SpliceDecl id)
Splice declaration (Includes quasi-quotes)

DocD DocDecl
Documentation comment declaration

RoleAnnotD (RoleAnnotDecl id)
Role annotation declaration

type LHsDecl id

Arguments

= Located (HsDecl id)
When in a list this may have

AnnKeywordId : AnnSemi

data HsDataDefn pass

Haskell Data type Definition

Constructors

HsDataDefn
Declares a data type or newtype, giving its constructors data/newtype T a = constrs data/newtype instance T [a] = constrs

Fields

dd_ND :: NewOrData

dd_ctxt :: LHsContext pass

Context

dd_cType :: Maybe (Located CType)

dd_kindSig :: Maybe (LHsKind pass)

Optional kind signature.

(Just k) for a GADT-style data, or data instance decl, with explicit kind sig

Always Nothing for H98-syntax decls

dd_cons :: [LConDecl pass]

Data constructors

For data T a = T1 | T2 a the LConDecls all have ConDeclH98. For data T a where { T1 :: T a } the LConDecls all have ConDeclGADT.

dd_derivs :: HsDeriving pass

Optional 'deriving' claues

type HsDeriving pass

Arguments

= Located [LHsDerivingClause pass]
The optional deriving clauses of a data declaration. Clauses is plural because one can specify multiple deriving clauses using the -XDerivingStrategies language extension.

The list of LHsDerivingClauses corresponds to exactly what the user requested to derive, in order. If no deriving clauses were specified, the list is empty.

Haskell Deriving clause

data HsDerivingClause pass

A single deriving clause of a data declaration.

AnnKeywordId : AnnDeriving, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

Constructors

HsDerivingClause

Fields

deriv_clause_strategy :: Maybe (Located DerivStrategy)

The user-specified strategy (if any) to use when deriving deriv_clause_tys.

deriv_clause_tys :: Located [LHsSigType pass]

The types to derive.

It uses LHsSigTypes because, with -XGeneralizedNewtypeDeriving, we can mention type variables that aren't bound by the datatype, e.g.

data T b = ... deriving (C [a])

should produce a derived instance for C [a] (T b).

type LHsDerivingClause pass = Located (HsDerivingClause pass)

Class or type declarations

data TyClDecl pass

A type or class declaration.

Constructors

FamDecl
type/data family T :: *->*

AnnKeywordId : AnnType, AnnData, AnnFamily,AnnDcolon, AnnWhere,AnnOpenP, AnnDcolon,AnnCloseP, AnnEqual,AnnRarrow, AnnVbar

Fields

tcdFam :: FamilyDecl pass

SynDecl
type declaration

AnnKeywordId : AnnType, AnnEqual,

Fields

tcdLName :: Located (IdP pass)

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdRhs :: LHsType pass

RHS of type declaration

tcdFVs :: PostRn pass NameSet

DataDecl
data declaration

AnnKeywordId : AnnData, AnnFamily, AnnNewType, AnnNewType,AnnDcolon AnnWhere,

Fields

tcdLName :: Located (IdP pass)

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdDataDefn :: HsDataDefn pass

tcdDataCusk :: PostRn pass Bool

does this have a CUSK?

tcdFVs :: PostRn pass NameSet

ClassDecl

AnnKeywordId : AnnClass, AnnWhere,AnnOpen, AnnClose

The tcdFDs will have AnnVbar, AnnComma AnnRarrow

Fields

tcdCtxt :: LHsContext pass

Context...

tcdLName :: Located (IdP pass)

Type constructor

tcdTyVars :: LHsQTyVars pass

Type variables; for an associated type these include outer binders

tcdFixity :: LexicalFixity

Fixity used in the declaration

tcdFDs :: [Located (FunDep (Located (IdP pass)))]

Functional deps

tcdSigs :: [LSig pass]

Methods' signatures

tcdMeths :: LHsBinds pass

Default methods

tcdATs :: [LFamilyDecl pass]

Associated types;

tcdATDefs :: [LTyFamDefltEqn pass]

Associated type defaults

tcdDocs :: [LDocDecl]

Haddock docs

tcdFVs :: PostRn pass NameSet

type LTyClDecl pass = Located (TyClDecl pass)

Located Declaration of a Type or Class

data TyClGroup pass

Type or Class Group

Constructors

TyClGroup

Fields

group_tyclds :: [LTyClDecl pass]

group_roles :: [LRoleAnnotDecl pass]

group_instds :: [LInstDecl pass]

mkTyClGroup :: [LTyClDecl pass] -> [LInstDecl pass] -> TyClGroup pass

emptyTyClGroup :: TyClGroup pass

tyClGroupTyClDecls :: [TyClGroup pass] -> [LTyClDecl pass]

tyClGroupInstDecls :: [TyClGroup pass] -> [LInstDecl pass]

tyClGroupRoleDecls :: [TyClGroup pass] -> [LRoleAnnotDecl pass]

isClassDecl :: TyClDecl pass -> Bool

type class

isDataDecl :: TyClDecl pass -> Bool

True = argument is a data/newtype declaration.

isSynDecl :: TyClDecl pass -> Bool

type or type instance declaration

tcdName :: TyClDecl pass -> IdP pass

isFamilyDecl :: TyClDecl pass -> Bool

type/data family declaration

isTypeFamilyDecl :: TyClDecl pass -> Bool

type family declaration

isDataFamilyDecl :: TyClDecl pass -> Bool

data family declaration

isOpenTypeFamilyInfo :: FamilyInfo pass -> Bool

open type family info

isClosedTypeFamilyInfo :: FamilyInfo pass -> Bool

closed type family info

tyFamInstDeclName :: TyFamInstDecl pass -> IdP pass

tyFamInstDeclLName :: TyFamInstDecl pass -> Located (IdP pass)

countTyClDecls :: [TyClDecl pass] -> (Int, Int, Int, Int, Int)

pprTyClDeclFlavour :: TyClDecl a -> SDoc

tyClDeclLName :: TyClDecl pass -> Located (IdP pass)

tyClDeclTyVars :: TyClDecl pass -> LHsQTyVars pass

hsDeclHasCusk :: TyClDecl GhcRn -> Bool

Does this declaration have a complete, user-supplied kind signature? See Note [Complete user-supplied kind signatures]

famDeclHasCusk

Arguments

:: Maybe Bool
if associated, does the enclosing class have a CUSK?

-> FamilyDecl pass

-> Bool

Does this family declaration have a complete, user-supplied kind signature?

data FamilyDecl pass

type Family Declaration

Constructors

FamilyDecl

AnnKeywordId : AnnType, AnnData, AnnFamily, AnnWhere, AnnOpenP, AnnDcolon, AnnCloseP, AnnEqual, AnnRarrow, AnnVbar

Fields

fdInfo :: FamilyInfo pass

fdLName :: Located (IdP pass)

fdTyVars :: LHsQTyVars pass

fdFixity :: LexicalFixity

fdResultSig :: LFamilyResultSig pass

fdInjectivityAnn :: Maybe (LInjectivityAnn pass)

type LFamilyDecl pass = Located (FamilyDecl pass)

Located type Family Declaration

Instance declarations

data InstDecl pass

Instance Declaration

Constructors

ClsInstD

Fields

cid_inst :: ClsInstDecl pass

DataFamInstD

Fields

dfid_inst :: DataFamInstDecl pass

TyFamInstD

Fields

tfid_inst :: TyFamInstDecl pass

type LInstDecl pass = Located (InstDecl pass)

Located Instance Declaration

data NewOrData

Constructors

NewType
newtype Blah ...

DataType
data Blah ...

data FamilyInfo pass

Constructors

DataFamily

OpenTypeFamily

ClosedTypeFamily (Maybe [LTyFamInstEqn pass])
Nothing if we're in an hs-boot file and the user said "type family Foo x where .."

newtype TyFamInstDecl pass

Type Family Instance Declaration

Constructors

TyFamInstDecl

AnnKeywordId : AnnType, AnnInstance,

Fields

tfid_eqn :: TyFamInstEqn pass

type LTyFamInstDecl pass = Located (TyFamInstDecl pass)

Located Type Family Instance Declaration

instDeclDataFamInsts :: [LInstDecl pass] -> [DataFamInstDecl pass]

newtype DataFamInstDecl pass

Data Family Instance Declaration

Constructors

DataFamInstDecl

AnnKeywordId : AnnData, AnnNewType,AnnInstance, AnnDcolon AnnWhere,AnnOpen, AnnClose

Fields

dfid_eqn :: FamInstEqn pass (HsDataDefn pass)

type LDataFamInstDecl pass = Located (DataFamInstDecl pass)

Located Data Family Instance Declaration

pprDataFamInstFlavour :: DataFamInstDecl pass -> SDoc

pprFamInstLHS :: (SourceTextX pass, OutputableBndrId pass) => Located (IdP pass) -> HsTyPats pass -> LexicalFixity -> HsContext pass -> Maybe (LHsKind pass) -> SDoc

type FamInstEqn pass rhs

Arguments

= HsImplicitBndrs pass (FamEqn pass (HsTyPats pass) rhs)
Here, the pats are type patterns (with kind and type bndrs). See Note [Family instance declaration binders]

Family Instance Equation

type LFamInstEqn pass rhs = Located (FamInstEqn pass rhs)

Located Family Instance Equation

data FamEqn pass pats rhs

Family Equation

One equation in a type family instance declaration, data family instance declaration, or type family default. See Note [Type family instance declarations in HsSyn] See Note [Family instance declaration binders]

Constructors

FamEqn

AnnKeywordId : AnnEqual

Fields

feqn_tycon :: Located (IdP pass)

feqn_pats :: pats

feqn_fixity :: LexicalFixity

Fixity used in the declaration

feqn_rhs :: rhs

type TyFamInstEqn pass = FamInstEqn pass (LHsType pass)

Type Family Instance Equation

type LTyFamInstEqn pass

Arguments

= Located (TyFamInstEqn pass)
May have AnnKeywordId : AnnSemi when in a list

Located Type Family Instance Equation

type TyFamDefltEqn pass = FamEqn pass (LHsQTyVars pass) (LHsType pass)

Type Family Default Equation

type LTyFamDefltEqn pass = Located (TyFamDefltEqn pass)

Located Type Family Default Equation

type HsTyPats pass = [LHsType pass]

Haskell Type Patterns

type LClsInstDecl pass = Located (ClsInstDecl pass)

Located Class Instance Declaration

data ClsInstDecl pass

Class Instance Declaration

Constructors

ClsInstDecl

AnnKeywordId : AnnInstance, AnnWhere, AnnOpen,AnnClose,

Fields

cid_poly_ty :: LHsSigType pass

cid_binds :: LHsBinds pass

cid_sigs :: [LSig pass]

cid_tyfam_insts :: [LTyFamInstDecl pass]

cid_datafam_insts :: [LDataFamInstDecl pass]

cid_overlap_mode :: Maybe (Located OverlapMode)

AnnKeywordId : AnnOpen, AnnClose,

Standalone deriving declarations

data DerivDecl pass

Deriving Declaration

Constructors

DerivDecl

Fields

deriv_type :: LHsSigType pass

deriv_strategy :: Maybe (Located DerivStrategy)

deriv_overlap_mode :: Maybe (Located OverlapMode)

AnnKeywordId : AnnDeriving, AnnInstance, AnnStock, AnnAnyClass, AnnNewtype, AnnOpen,AnnClose

type LDerivDecl pass = Located (DerivDecl pass)

Located Deriving Declaration

RULE declarations

type LRuleDecls pass = Located (RuleDecls pass)

Located Rule Declarations

data RuleDecls pass

Rule Declarations

Constructors

HsRules

Fields

rds_src :: SourceText

rds_rules :: [LRuleDecl pass]

data RuleDecl pass

Rule Declaration

Constructors

HsRule (Located (SourceText, RuleName)) Activation [LRuleBndr pass] (Located (HsExpr pass)) (PostRn pass NameSet) (Located (HsExpr pass)) (PostRn pass NameSet)

AnnKeywordId : AnnOpen,AnnTilde, AnnVal, AnnClose, AnnForall,AnnDot, AnnEqual,

type LRuleDecl pass = Located (RuleDecl pass)

Located Rule Declaration

data RuleBndr pass

Rule Binder

Constructors

RuleBndr (Located (IdP pass))

RuleBndrSig (Located (IdP pass)) (LHsSigWcType pass)

AnnKeywordId : AnnOpen, AnnDcolon,AnnClose

type LRuleBndr pass = Located (RuleBndr pass)

Located Rule Binder

collectRuleBndrSigTys :: [RuleBndr pass] -> [LHsSigWcType pass]

flattenRuleDecls :: [LRuleDecls pass] -> [LRuleDecl pass]

pprFullRuleName :: Located (SourceText, RuleName) -> SDoc

VECTORISE declarations

data VectDecl pass

Vectorise Declaration

Constructors

HsVect SourceText (Located (IdP pass)) (LHsExpr pass)

AnnKeywordId : AnnOpen, AnnEqual,AnnClose

HsNoVect SourceText (Located (IdP pass))

AnnKeywordId : AnnOpen, AnnClose

HsVectTypeIn SourceText Bool (Located (IdP pass)) (Maybe (Located (IdP pass)))

AnnKeywordId : AnnOpen, AnnType,AnnClose, AnnEqual

HsVectTypeOut Bool TyCon (Maybe TyCon)

HsVectClassIn SourceText (Located (IdP pass))

AnnKeywordId : AnnOpen, AnnClass,AnnClose,

HsVectClassOut Class

HsVectInstIn (LHsSigType pass)

HsVectInstOut ClsInst

type LVectDecl pass = Located (VectDecl pass)

Located Vectorise Declaration

lvectDeclName :: NamedThing (IdP pass) => LVectDecl pass -> Name

lvectInstDecl :: LVectDecl pass -> Bool

default declarations

data DefaultDecl pass

Default Declaration

Constructors

DefaultDecl [LHsType pass]

AnnKeywordIds : AnnDefault, AnnOpen,AnnClose

type LDefaultDecl pass = Located (DefaultDecl pass)

Located Default Declaration

Template haskell declaration splice

data SpliceExplicitFlag

Constructors

ExplicitSplice
= $(f x y)

ImplicitSplice
= f x y, i.e. a naked top level expression

data SpliceDecl id

Splice Declaration

Constructors

SpliceDecl (Located (HsSplice id)) SpliceExplicitFlag

type LSpliceDecl pass = Located (SpliceDecl pass)

Located Splice Declaration

Foreign function interface declarations

data ForeignDecl pass

Foreign Declaration

Constructors

ForeignImport

Fields

fd_name :: Located (IdP pass)

fd_sig_ty :: LHsSigType pass

fd_co :: PostTc pass Coercion

fd_fi :: ForeignImport

ForeignExport

AnnKeywordId : AnnForeign, AnnImport,AnnExport, AnnDcolon

Fields

fd_name :: Located (IdP pass)

fd_sig_ty :: LHsSigType pass

fd_co :: PostTc pass Coercion

fd_fe :: ForeignExport

type LForeignDecl pass = Located (ForeignDecl pass)

Located Foreign Declaration

data ForeignImport

Constructors

CImport (Located CCallConv) (Located Safety) (Maybe Header) CImportSpec (Located SourceText)

data ForeignExport

Constructors

CExport (Located CExportSpec) (Located SourceText)

noForeignImportCoercionYet :: PlaceHolder

noForeignExportCoercionYet :: PlaceHolder

data CImportSpec

Constructors

CLabel CLabelString

CFunction CCallTarget

CWrapper

Data-constructor declarations

data ConDecl pass

data T b = forall a. Eq a => MkT a b
  MkT :: forall b a. Eq a => MkT a b

data T b where
     MkT1 :: Int -> T Int

data T = Int MkT Int
       | MkT2

data T a where
     Int MkT Int :: T Int

AnnKeywordIds : AnnOpen, AnnDotdot,AnnCLose, AnnEqual,AnnVbar, AnnDarrow,AnnDarrow, AnnForall,AnnDot

data Constructor Declaration

Constructors

ConDeclGADT

Fields

con_names :: [Located (IdP pass)]

con_type :: LHsSigType pass

The type after the ‘::’

con_doc :: Maybe LHsDocString

A possible Haddock comment.

ConDeclH98

Fields

con_name :: Located (IdP pass)

con_qvars :: Maybe (LHsQTyVars pass)

con_cxt :: Maybe (LHsContext pass)

User-written context (if any)

con_details :: HsConDeclDetails pass

Arguments

con_doc :: Maybe LHsDocString

A possible Haddock comment.

type LConDecl pass

Arguments

= Located (ConDecl pass)
May have AnnKeywordId : AnnSemi when in a GADT constructor list

Located data Constructor Declaration

type HsConDeclDetails pass = HsConDetails (LBangType pass) (Located [LConDeclField pass])

Haskell data Constructor Declaration Details

hsConDeclArgTys :: HsConDeclDetails pass -> [LBangType pass]

getConNames :: ConDecl pass -> [Located (IdP pass)]

getConDetails :: ConDecl pass -> HsConDeclDetails pass

gadtDeclDetails :: LHsSigType pass -> (HsConDeclDetails pass, LHsType pass, LHsContext pass, [LHsTyVarBndr pass])

Document comments

data DocDecl

Documentation comment Declaration

Constructors

DocCommentNext HsDocString

DocCommentPrev HsDocString

DocCommentNamed String HsDocString

DocGroup Int HsDocString

type LDocDecl = Located DocDecl

Located Documentation comment Declaration

docDeclDoc :: DocDecl -> HsDocString

Deprecations

data WarnDecl pass

Warning pragma Declaration

Constructors

Warning [Located (IdP pass)] WarningTxt

type LWarnDecl pass = Located (WarnDecl pass)

Located Warning pragma Declaration

data WarnDecls pass

Warning pragma Declarations

Constructors

Warnings

Fields

wd_src :: SourceText

wd_warnings :: [LWarnDecl pass]

type LWarnDecls pass = Located (WarnDecls pass)

Located Warning Declarations

Annotations

data AnnDecl pass

Annotation Declaration

Constructors

HsAnnotation SourceText (AnnProvenance (IdP pass)) (Located (HsExpr pass))

AnnKeywordId : AnnOpen, AnnType AnnModule AnnClose

type LAnnDecl pass = Located (AnnDecl pass)

Located Annotation Declaration

data AnnProvenance name

Annotation Provenance

Constructors

ValueAnnProvenance (Located name)

TypeAnnProvenance (Located name)

ModuleAnnProvenance

annProvenanceName_maybe :: AnnProvenance name -> Maybe name

Role annotations

data RoleAnnotDecl pass

Role Annotation Declaration

Constructors

RoleAnnotDecl (Located (IdP pass)) [Located (Maybe Role)]

AnnKeywordId : AnnType, AnnRole

type LRoleAnnotDecl pass = Located (RoleAnnotDecl pass)

Located Role Annotation Declaration

roleAnnotDeclName :: RoleAnnotDecl pass -> IdP pass

Injective type families

data FamilyResultSig pass

type Family Result Signature

Constructors

NoSig

AnnKeywordId :

KindSig (LHsKind pass)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP

TyVarSig (LHsTyVarBndr pass)

AnnKeywordId : AnnOpenP,AnnDcolon, AnnCloseP, AnnEqual

type LFamilyResultSig pass = Located (FamilyResultSig pass)

Located type Family Result Signature

data InjectivityAnn pass

If the user supplied an injectivity annotation it is represented using InjectivityAnn. At the moment this is a single injectivity condition - see Note [Injectivity annotation]. `Located name` stores the LHS of injectivity condition. `[Located name]` stores the RHS of injectivity condition. Example:

type family Foo a b c = r | r -> a c where ...

This will be represented as "InjectivityAnn r [a, c]"

Constructors

InjectivityAnn (Located (IdP pass)) [Located (IdP pass)]

AnnKeywordId : AnnRarrow, AnnVbar

type LInjectivityAnn pass = Located (InjectivityAnn pass)

Located Injectivity Annotation

resultVariableName :: FamilyResultSig a -> Maybe (IdP a)

Maybe return name of the result type variable

Grouping

data HsGroup id

Haskell Group

A HsDecl is categorised into a HsGroup before being fed to the renamer.

Constructors

HsGroup

Fields

hs_valds :: HsValBinds id

hs_splcds :: [LSpliceDecl id]

hs_tyclds :: [TyClGroup id]

hs_derivds :: [LDerivDecl id]

hs_fixds :: [LFixitySig id]

hs_defds :: [LDefaultDecl id]

hs_fords :: [LForeignDecl id]

hs_warnds :: [LWarnDecls id]

hs_annds :: [LAnnDecl id]

hs_ruleds :: [LRuleDecls id]

hs_vects :: [LVectDecl id]

hs_docs :: [LDocDecl]

emptyRdrGroup :: HsGroup a

emptyRnGroup :: HsGroup a

appendGroups :: HsGroup a -> HsGroup a -> HsGroup a

hsGroupInstDecls :: HsGroup id -> [LInstDecl id]