HsDecls

Description

Abstract syntax of global declarations.

Definitions for: TyDecl and ConDecl, ClassDecl, InstDecl, DefaultDecl and ForeignDecl.

Toplevel declarations

data HsDecl id

A Haskell Declaration

Constructors

TyClD (TyClDecl id)
A type or class declaration.

InstD (InstDecl id)
An instance declaration.

DerivD (DerivDecl id)

ValD (HsBind id)

SigD (Sig id)

DefD (DefaultDecl id)

ForD (ForeignDecl id)

WarningD (WarnDecl id)

AnnD (AnnDecl id)

RuleD (RuleDecl id)

VectD (VectDecl id)

SpliceD (SpliceDecl id)

DocD DocDecl

QuasiQuoteD (HsQuasiQuote id)

type LHsDecl id = Located (HsDecl id)

Class or type declarations

data TyClDecl name

A type or class declaration.

Constructors

ForeignType

Fields

tcdLName :: Located name

Name of the class

type constructor

Type constructor

tcdExtName :: Maybe FastString

TyFamily
type/data family T :: *->*

Fields

tcdFlavour :: FamilyFlavour

tcdLName :: Located name

Name of the class

type constructor

Type constructor

tcdTyVars :: [LHsTyVarBndr name]

Class type variables

type variables

Type variables

tcdKind :: Maybe (LHsKind name)

TyData
Declares a data type or newtype, giving its construcors data/newtype T a = constrs data/newtype instance T [a] = constrs

Fields

tcdND :: NewOrData

tcdCtxt :: LHsContext name

Context...

Context

tcdLName :: Located name

Name of the class

type constructor

Type constructor

tcdTyVars :: [LHsTyVarBndr name]

Class type variables

type variables

Type variables

tcdTyPats :: Maybe [LHsType name]

Type patterns See Note [tcdTyVars and tcdTyPats]

Type patterns. See Note [tcdTyVars and tcdTyPats]

tcdKindSig :: Maybe (LHsKind name)

Optional kind signature.

(Just k) for a GADT-style data, or data instance decl with explicit kind sig

tcdCons :: [LConDecl name]

Data constructors

For data T a = T1 | T2 a the LConDecls all have ResTyH98. For data T a where { T1 :: T a } the LConDecls all have ResTyGADT.

tcdDerivs :: Maybe [LHsType name]

Derivings; Nothing => not specified, Just [] => derive exactly what is asked

These types must be of form forall ab. C ty1 ty2 Typically the foralls and ty args are empty, but they are non-empty for the newtype-deriving case

TySynonym

Fields

tcdLName :: Located name

Name of the class

type constructor

Type constructor

tcdTyVars :: [LHsTyVarBndr name]

Class type variables

type variables

Type variables

tcdTyPats :: Maybe [LHsType name]

Type patterns See Note [tcdTyVars and tcdTyPats]

Type patterns. See Note [tcdTyVars and tcdTyPats]

tcdSynRhs :: LHsType name

synonym expansion

ClassDecl

Fields

tcdCtxt :: LHsContext name

Context...

Context

tcdLName :: Located name

Name of the class

type constructor

Type constructor

tcdTyVars :: [LHsTyVarBndr name]

Class type variables

type variables

Type variables

tcdFDs :: [Located (FunDep name)]

Functional deps

tcdSigs :: [LSig name]

Methods' signatures

tcdMeths :: LHsBinds name

Default methods

tcdATs :: [LTyClDecl name]

Associated types; ie only TyFamily

tcdATDefs :: [LTyClDecl name]

Associated type defaults; ie only TySynonym

tcdDocs :: [LDocDecl]

Haddock docs

type LTyClDecl name = Located (TyClDecl name)

type TyClGroup name = [LTyClDecl name]

isClassDecl :: TyClDecl name -> Bool

type class

isSynDecl :: TyClDecl name -> Bool

vanilla Haskell type synonym (ie, not a type instance)

isDataDecl :: TyClDecl name -> Bool

True = argument is a data/newtype or data/newtype instance declaration.

isTypeDecl :: TyClDecl name -> Bool

type or type instance declaration

isFamilyDecl :: TyClDecl name -> Bool

type family declaration

isFamInstDecl :: TyClDecl name -> Bool

family instance (types, newtypes, and data types)

tcdName :: TyClDecl name -> name

tyClDeclTyVars :: TyClDecl name -> [LHsTyVarBndr name]

countTyClDecls :: [TyClDecl name] -> (Int, Int, Int, Int, Int, Int)

Instance declarations

data InstDecl name

Constructors

InstDecl (LHsType name) (LHsBinds name) [LSig name] [LTyClDecl name]

type LInstDecl name = Located (InstDecl name)

data NewOrData

Constructors

NewType
newtype Blah ...

DataType
data Blah ...

data FamilyFlavour

Constructors

TypeFamily
type family ...

DataFamily
data family ...

instDeclATs :: [LInstDecl name] -> [LTyClDecl name]

Standalone deriving declarations

data DerivDecl name

Constructors

DerivDecl

Fields

deriv_type :: LHsType name

type LDerivDecl name = Located (DerivDecl name)

RULE declarations

data RuleDecl name

Constructors

HsRule RuleName Activation [RuleBndr name] (Located (HsExpr name)) NameSet (Located (HsExpr name)) NameSet

type LRuleDecl name = Located (RuleDecl name)

data RuleBndr name

Constructors

RuleBndr (Located name)

RuleBndrSig (Located name) (LHsType name)

collectRuleBndrSigTys :: [RuleBndr name] -> [LHsType name]

VECTORISE declarations

data VectDecl name

Constructors

HsVect (Located name) (Maybe (LHsExpr name))

HsNoVect (Located name)

HsVectTypeIn Bool (Located name) (Maybe (Located name))

HsVectTypeOut Bool TyCon (Maybe TyCon)

HsVectClassIn (Located name)

HsVectClassOut Class

HsVectInstIn (LHsType name)

HsVectInstOut Instance

type LVectDecl name = Located (VectDecl name)

lvectDeclName :: NamedThing name => LVectDecl name -> Name

lvectInstDecl :: LVectDecl name -> Bool

default declarations

data DefaultDecl name

Constructors

DefaultDecl [LHsType name]

type LDefaultDecl name = Located (DefaultDecl name)

Top-level template haskell splice

data SpliceDecl id

Constructors

SpliceDecl (Located (HsExpr id)) HsExplicitFlag

Foreign function interface declarations

data ForeignDecl name

Constructors

ForeignImport (Located name) (LHsType name) Coercion ForeignImport

ForeignExport (Located name) (LHsType name) Coercion ForeignExport

type LForeignDecl name = Located (ForeignDecl name)

data ForeignImport

Constructors

CImport CCallConv Safety FastString CImportSpec

data ForeignExport

Constructors

CExport CExportSpec

noForeignImportCoercionYet :: Coercion

noForeignExportCoercionYet :: Coercion

data CImportSpec

Constructors

CLabel CLabelString

CFunction CCallTarget

CWrapper

Data-constructor declarations

data ConDecl name

Constructors

ConDecl

Fields

con_name :: Located name

Constructor name. This is used for the DataCon itself, and for the user-callable wrapper Id.

con_explicit :: HsExplicitFlag

Is there an user-written forall? (cf. HsForAllTy)

con_qvars :: [LHsTyVarBndr name]

Type variables. Depending on con_res this describes the following entities

ResTyH98: the constructor's *existential* type variables - ResTyGADT: *all* the constructor's quantified type variables

If con_explicit is Implicit, then con_qvars is irrelevant until after renaming.

con_cxt :: LHsContext name

The context. This does not include the "stupid theta" which lives only in the TyData decl.

con_details :: HsConDeclDetails name

The main payload

con_res :: ResType name

Result type of the constructor

con_doc :: Maybe LHsDocString

A possible Haddock comment.

con_old_rec :: Bool

TEMPORARY field; True = user has employed now-deprecated syntax for GADT-style record decl C { blah } :: T a b Remove this when we no longer parse this stuff, and hence do not need to report decprecated use

type LConDecl name = Located (ConDecl name)

data ResType name

Constructors

ResTyH98

ResTyGADT (LHsType name)

type HsConDeclDetails name = HsConDetails (LBangType name) [ConDeclField name]

hsConDeclArgTys :: HsConDeclDetails name -> [LBangType name]

Document comments

data DocDecl

Constructors

DocCommentNext HsDocString

DocCommentPrev HsDocString

DocCommentNamed String HsDocString

DocGroup Int HsDocString

type LDocDecl = Located DocDecl

docDeclDoc :: DocDecl -> HsDocString

Deprecations

data WarnDecl name

Constructors

Warning name WarningTxt

type LWarnDecl name = Located (WarnDecl name)

Annotations

data AnnDecl name

Constructors

HsAnnotation (AnnProvenance name) (Located (HsExpr name))

type LAnnDecl name = Located (AnnDecl name)

data AnnProvenance name

Constructors

ValueAnnProvenance name

TypeAnnProvenance name

ModuleAnnProvenance

annProvenanceName_maybe :: AnnProvenance name -> Maybe name

modifyAnnProvenanceNameM :: Monad m => (before -> m after) -> AnnProvenance before -> m (AnnProvenance after)

Grouping

data HsGroup id

A HsDecl is categorised into a HsGroup before being fed to the renamer.

Constructors

HsGroup

Fields

hs_valds :: HsValBinds id

hs_tyclds :: [[LTyClDecl id]]

hs_instds :: [LInstDecl id]

hs_derivds :: [LDerivDecl id]

hs_fixds :: [LFixitySig id]

hs_defds :: [LDefaultDecl id]

hs_fords :: [LForeignDecl id]

hs_warnds :: [LWarnDecl id]

hs_annds :: [LAnnDecl id]

hs_ruleds :: [LRuleDecl id]

hs_vects :: [LVectDecl id]

hs_docs :: [LDocDecl]

emptyRdrGroup :: HsGroup a

emptyRnGroup :: HsGroup a

appendGroups :: HsGroup a -> HsGroup a -> HsGroup a