HsExpr

Description

Abstract Haskell syntax for expressions.

Expressions proper

type LHsExpr id

Arguments

= Located (HsExpr id)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Expression

type PostTcExpr = HsExpr Id

Post-Type checking Expression

PostTcExpr is an evidence expression attached to the syntax tree by the type checker (c.f. postTcType).

type PostTcTable = [(Name, PostTcExpr)]

Post-Type checking Table

We use a PostTcTable where there are a bunch of pieces of evidence, more than is convenient to keep individually.

noPostTcExpr :: PostTcExpr

noPostTcTable :: PostTcTable

data SyntaxExpr id

Syntax Expression

SyntaxExpr is like PostTcExpr, but it's filled in a little earlier, by the renamer. It's used for rebindable syntax.

E.g. (>>=) is filled in before the renamer by the appropriate Name for (>>=), and then instantiated by the type checker with its type args etc

This should desugar to

syn_res_wrap $ syn_expr (syn_arg_wraps[0] arg0)
                        (syn_arg_wraps[1] arg1) ...

where the actual arguments come from elsewhere in the AST. This could be defined using PostRn and PostTc and such, but it's harder to get it all to work out that way. (noSyntaxExpr is hard to write, for example.)

Constructors

SyntaxExpr

Fields

syn_expr :: HsExpr id

syn_arg_wraps :: [HsWrapper]

syn_res_wrap :: HsWrapper

noExpr :: HsExpr id

This is used for rebindable-syntax pieces that are too polymorphic for tcSyntaxOp (trS_fmap and the mzip in ParStmt)

noSyntaxExpr :: SyntaxExpr id

mkRnSyntaxExpr :: Name -> SyntaxExpr Name

Make a 'SyntaxExpr Name' (the "rn" is because this is used in the renamer), missing its HsWrappers.

type CmdSyntaxTable id = [(Name, HsExpr id)]

Command Syntax Table (for Arrow syntax)

data UnboundVar

An unbound variable; used for treating out-of-scope variables as expression holes

Constructors

OutOfScope OccName GlobalRdrEnv
An (unqualified) out-of-scope variable, together with the GlobalRdrEnv with respect to which it is unbound

TrueExprHole OccName
A "true" expression hole (_ or _x)

unboundVarOcc :: UnboundVar -> OccName

data HsExpr id

A Haskell expression.

Constructors

HsVar (Located id)
Variable

HsUnboundVar UnboundVar
Unbound variable; also used for "holes" (_ or _x). Turned from HsVar to HsUnboundVar by the renamer, when it finds an out-of-scope variable or hole. Turned into HsVar by type checker, to support deferred type errors.

HsConLikeOut ConLike
After typechecker only; must be different HsVar for pretty printing

HsRecFld (AmbiguousFieldOcc id)
Variable pointing to record selector Not in use after typechecking

HsOverLabel (Maybe id) FastString
Overloaded label (Note [Overloaded labels] in GHC.OverloadedLabels) Just id means RebindableSyntax is in use, and gives the id of the in-scope fromLabel. NB: Not in use after typechecking

HsIPVar HsIPName
Implicit parameter (not in use after typechecking)

HsOverLit (HsOverLit id)
Overloaded literals

HsLit HsLit
Simple (non-overloaded) literals

HsLam (MatchGroup id (LHsExpr id))
Lambda abstraction. Currently always a single match

AnnKeywordId : AnnLam, AnnRarrow,

HsLamCase (MatchGroup id (LHsExpr id))
Lambda-case

AnnKeywordId : AnnLam, AnnCase,AnnOpen, AnnClose

HsApp (LHsExpr id) (LHsExpr id)
Application

HsAppType (LHsExpr id) (LHsWcType id)
Visible type application

Explicit type argument; e.g f @Int x y NB: Has wildcards, but no implicit quantification

AnnKeywordId : AnnAt,

HsAppTypeOut (LHsExpr id) (LHsWcType Name)

OpApp (LHsExpr id) (LHsExpr id) (PostRn id Fixity) (LHsExpr id)
Operator applications: NB Bracketed ops such as (+) come out as Vars.

NegApp (LHsExpr id) (SyntaxExpr id)
Negation operator. Contains the negated expression and the name of negate

AnnKeywordId : AnnMinus

HsPar (LHsExpr id)
Parenthesised expr; see Note [Parens in HsSyn]

SectionL (LHsExpr id) (LHsExpr id)

SectionR (LHsExpr id) (LHsExpr id)

ExplicitTuple [LHsTupArg id] Boxity
Used for explicit tuples and sections thereof

AnnKeywordId : AnnOpen, AnnClose

ExplicitSum ConTag Arity (LHsExpr id) (PostTc id [Type])
Used for unboxed sum types

AnnKeywordId : AnnOpen '(#', AnnVbar, AnnClose '#)',

There will be multiple AnnVbar, (1 - alternative) before the expression, (arity - alternative) after it

HsCase (LHsExpr id) (MatchGroup id (LHsExpr id))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsIf (Maybe (SyntaxExpr id)) (LHsExpr id) (LHsExpr id) (LHsExpr id)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsMultiIf (PostTc id Type) [LGRHS id (LHsExpr id)]
Multi-way if

AnnKeywordId : AnnIf AnnOpen,AnnClose,

HsLet (LHsLocalBinds id) (LHsExpr id)
let(rec)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsDo (HsStmtContext Name) (Located [ExprLStmt id]) (PostTc id Type)

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

ExplicitList (PostTc id Type) (Maybe (SyntaxExpr id)) [LHsExpr id]
Syntactic list: [a,b,c,...]

AnnKeywordId : AnnOpen '[', AnnClose ']'

ExplicitPArr (PostTc id Type) [LHsExpr id]
Syntactic parallel array: [:e1, ..., en:]

AnnKeywordId : AnnOpen '[:', AnnDotdot,AnnComma, AnnVbar AnnClose ':]'

RecordCon
Record construction

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rcon_con_name :: Located id

rcon_con_like :: PostTc id ConLike

rcon_con_expr :: PostTcExpr

rcon_flds :: HsRecordBinds id

RecordUpd
Record update

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rupd_expr :: LHsExpr id

rupd_flds :: [LHsRecUpdField id]

rupd_cons :: PostTc id [ConLike]

rupd_in_tys :: PostTc id [Type]

rupd_out_tys :: PostTc id [Type]

rupd_wrap :: PostTc id HsWrapper

ExprWithTySig (LHsExpr id) (LHsSigWcType id)
Expression with an explicit type signature. e :: type

AnnKeywordId : AnnDcolon

ExprWithTySigOut (LHsExpr id) (LHsSigWcType Name)

ArithSeq PostTcExpr (Maybe (SyntaxExpr id)) (ArithSeqInfo id)
Arithmetic sequence

AnnKeywordId : AnnOpen '[', AnnComma,AnnDotdot, AnnClose ']'

PArrSeq PostTcExpr (ArithSeqInfo id)
Arithmetic sequence for parallel array

[:e1..e2:] or [:e1, e2..e3:]

AnnKeywordId : AnnOpen '[:', AnnComma,AnnDotdot, AnnVbar, AnnClose ':]'

HsSCC SourceText StringLiteral (LHsExpr id)

AnnKeywordId : AnnOpen '{-# SCC', AnnVal or AnnValStr, AnnClose '#-}'

HsCoreAnn SourceText StringLiteral (LHsExpr id)

AnnKeywordId : AnnOpen '{-# CORE', AnnVal, AnnClose '#-}'

HsBracket (HsBracket id)

AnnKeywordId : AnnOpen, AnnOpenE,AnnOpenEQ, AnnClose,AnnCloseQ

HsRnBracketOut (HsBracket Name) [PendingRnSplice]

HsTcBracketOut (HsBracket Name) [PendingTcSplice]

HsSpliceE (HsSplice id)

AnnKeywordId : AnnOpen, AnnClose

HsProc (LPat id) (LHsCmdTop id)
proc notation for Arrows

AnnKeywordId : AnnProc, AnnRarrow

HsStatic (PostRn id NameSet) (LHsExpr id)

AnnKeywordId : AnnStatic,

HsArrApp (LHsExpr id) (LHsExpr id) (PostTc id Type) HsArrAppType Bool

AnnKeywordId : Annlarrowtail, Annrarrowtail,AnnLarrowtail, AnnRarrowtail

HsArrForm (LHsExpr id) (Maybe Fixity) [LHsCmdTop id]

AnnKeywordId : AnnOpenB '(|', AnnCloseB '|)'

HsTick (Tickish id) (LHsExpr id)

HsBinTick Int Int (LHsExpr id)

HsTickPragma SourceText (StringLiteral, (Int, Int), (Int, Int)) ((SourceText, SourceText), (SourceText, SourceText)) (LHsExpr id)

AnnKeywordId : AnnOpen, AnnOpen '{-# GENERATED', AnnVal,AnnVal, AnnColon,AnnVal, AnnMinus, AnnVal,AnnColon, AnnVal, AnnClose '#-}'

EWildPat

EAsPat (Located id) (LHsExpr id)

AnnKeywordId : AnnAt

EViewPat (LHsExpr id) (LHsExpr id)

AnnKeywordId : AnnRarrow

ELazyPat (LHsExpr id)

AnnKeywordId : AnnTilde

HsWrap HsWrapper (HsExpr id)

type LHsTupArg id = Located (HsTupArg id)

Located Haskell Tuple Argument

HsTupArg is used for tuple sections (,a,) is represented by ExplicitTuple [Missing ty1, Present a, Missing ty3] Which in turn stands for (x:ty1 y:ty2. (x,a,y))

data HsTupArg id

AnnKeywordId : AnnComma

Haskell Tuple Argument

Constructors

Present (LHsExpr id)
The argument

Missing (PostTc id Type)
The argument is missing, but this is its type

tupArgPresent :: LHsTupArg id -> Bool

pprLExpr :: OutputableBndrId id => LHsExpr id -> SDoc

pprExpr :: OutputableBndrId id => HsExpr id -> SDoc

isQuietHsExpr :: HsExpr id -> Bool

pprBinds :: (OutputableBndrId idL, OutputableBndrId idR) => HsLocalBindsLR idL idR -> SDoc

ppr_lexpr :: OutputableBndrId id => LHsExpr id -> SDoc

ppr_expr :: forall id. OutputableBndrId id => HsExpr id -> SDoc

data LHsWcTypeX

Located Haskell Wildcard Type Expression

Constructors

OutputableBndrId id => LHsWcTypeX (LHsWcType id)

ppr_apps :: OutputableBndrId id => HsExpr id -> [Either (LHsExpr id) LHsWcTypeX] -> SDoc

pprExternalSrcLoc :: (StringLiteral, (Int, Int), (Int, Int)) -> SDoc

pprDebugParendExpr :: OutputableBndrId id => LHsExpr id -> SDoc

pprParendLExpr :: OutputableBndrId id => LHsExpr id -> SDoc

pprParendExpr :: OutputableBndrId id => HsExpr id -> SDoc

hsExprNeedsParens :: HsExpr id -> Bool

isAtomicHsExpr :: HsExpr id -> Bool

type LHsCmd id = Located (HsCmd id)

Located Haskell Command (for arrow syntax)

data HsCmd id

Haskell Command (e.g. a "statement" in an Arrow proc block)

Constructors

HsCmdArrApp (LHsExpr id) (LHsExpr id) (PostTc id Type) HsArrAppType Bool

AnnKeywordId : Annlarrowtail, Annrarrowtail,AnnLarrowtail, AnnRarrowtail

HsCmdArrForm (LHsExpr id) LexicalFixity (Maybe Fixity) [LHsCmdTop id]

AnnKeywordId : AnnOpenB '(|', AnnCloseB '|)'

HsCmdApp (LHsCmd id) (LHsExpr id)

HsCmdLam (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnLam, AnnRarrow,

HsCmdPar (LHsCmd id)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsCmdCase (LHsExpr id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsCmdIf (Maybe (SyntaxExpr id)) (LHsExpr id) (LHsCmd id) (LHsCmd id)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsCmdLet (LHsLocalBinds id) (LHsCmd id)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsCmdDo (Located [CmdLStmt id]) (PostTc id Type)

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

HsCmdWrap HsWrapper (HsCmd id)

data HsArrAppType

Haskell Array Application Type

Constructors

HsHigherOrderApp

HsFirstOrderApp

type LHsCmdTop id = Located (HsCmdTop id)

Top-level command, introducing a new arrow. This may occur inside a proc (where the stack is empty) or as an argument of a command-forming operator.

Located Haskell Top-level Command

data HsCmdTop id

Haskell Top-level Command

Constructors

HsCmdTop (LHsCmd id) (PostTc id Type) (PostTc id Type) (CmdSyntaxTable id)

pprLCmd :: OutputableBndrId id => LHsCmd id -> SDoc

pprCmd :: OutputableBndrId id => HsCmd id -> SDoc

isQuietHsCmd :: HsCmd id -> Bool

ppr_lcmd :: OutputableBndrId id => LHsCmd id -> SDoc

ppr_cmd :: forall id. OutputableBndrId id => HsCmd id -> SDoc

pprCmdArg :: OutputableBndrId id => HsCmdTop id -> SDoc

type HsRecordBinds id = HsRecFields id (LHsExpr id)

Haskell Record Bindings

data MatchGroup id body

Constructors

MG

Fields

mg_alts :: Located [LMatch id body]

mg_arg_tys :: [PostTc id Type]

mg_res_ty :: PostTc id Type

mg_origin :: Origin

type LMatch id body = Located (Match id body)

Located Match

May have AnnKeywordId : AnnSemi when in a list

data Match id body

Constructors

Match

Fields

m_ctxt :: HsMatchContext (NameOrRdrName id)

m_pats :: [LPat id]

m_type :: Maybe (LHsType id)

m_grhss :: GRHSs id body

isInfixMatch :: Match id body -> Bool

isEmptyMatchGroup :: MatchGroup id body -> Bool

isSingletonMatchGroup :: [LMatch id body] -> Bool

Is there only one RHS in this list of matches?

matchGroupArity :: MatchGroup id body -> Arity

hsLMatchPats :: LMatch id body -> [LPat id]

data GRHSs id body

Guarded Right-Hand Sides

GRHSs are used both for pattern bindings and for Matches

AnnKeywordId : AnnVbar, AnnEqual,AnnWhere, AnnOpen,AnnClose AnnRarrow,AnnSemi

Constructors

GRHSs

Fields

grhssGRHSs :: [LGRHS id body]

Guarded RHSs

grhssLocalBinds :: LHsLocalBinds id

The where clause

type LGRHS id body = Located (GRHS id body)

Located Guarded Right-Hand Side

data GRHS id body

Guarded Right Hand Side.

Constructors

GRHS [GuardLStmt id] body

pprMatches :: (OutputableBndrId idR, Outputable body) => MatchGroup idR body -> SDoc

pprFunBind :: (OutputableBndrId idR, Outputable body) => MatchGroup idR body -> SDoc

pprPatBind :: forall bndr id body. (OutputableBndrId bndr, OutputableBndrId id, Outputable body) => LPat bndr -> GRHSs id body -> SDoc

pprMatch :: (OutputableBndrId idR, Outputable body) => Match idR body -> SDoc

pprGRHSs :: (OutputableBndrId idR, Outputable body) => HsMatchContext idL -> GRHSs idR body -> SDoc

pprGRHS :: (OutputableBndrId idR, Outputable body) => HsMatchContext idL -> GRHS idR body -> SDoc

pp_rhs :: Outputable body => HsMatchContext idL -> body -> SDoc

type LStmt id body = Located (StmtLR id id body)

Located do block Statement

type LStmtLR idL idR body = Located (StmtLR idL idR body)

Located Statement with separate Left and Right id's

type Stmt id body = StmtLR id id body

do block Statement

type CmdLStmt id = LStmt id (LHsCmd id)

Command Located Statement

type CmdStmt id = Stmt id (LHsCmd id)

Command Statement

type ExprLStmt id = LStmt id (LHsExpr id)

Expression Located Statement

type ExprStmt id = Stmt id (LHsExpr id)

Expression Statement

type GuardLStmt id = LStmt id (LHsExpr id)

Guard Located Statement

type GuardStmt id = Stmt id (LHsExpr id)

Guard Statement

type GhciLStmt id = LStmt id (LHsExpr id)

Ghci Located Statemnt

type GhciStmt id = Stmt id (LHsExpr id)

Ghci Statement

data StmtLR idL idR body

API Annotations when in qualifier lists or guards - AnnKeywordId : AnnVbar, AnnComma,AnnThen, AnnBy,AnnBy, AnnGroup,AnnUsing

Constructors

LastStmt body Bool (SyntaxExpr idR)

BindStmt (LPat idL) body (SyntaxExpr idR) (SyntaxExpr idR) (PostTc idR Type)

ApplicativeStmt [(SyntaxExpr idR, ApplicativeArg idL idR)] (Maybe (SyntaxExpr idR)) (PostTc idR Type)
ApplicativeStmt represents an applicative expression built with $ and *. It is generated by the renamer, and is desugared into the appropriate applicative expression by the desugarer, but it is intended to be invisible in error messages.

For full details, see Note [ApplicativeDo] in RnExpr

BodyStmt body (SyntaxExpr idR) (SyntaxExpr idR) (PostTc idR Type)

LetStmt (LHsLocalBindsLR idL idR)

AnnKeywordId : AnnLet AnnOpen '{',AnnClose '}',

ParStmt [ParStmtBlock idL idR] (HsExpr idR) (SyntaxExpr idR) (PostTc idR Type)

TransStmt

Fields

trS_form :: TransForm

trS_stmts :: [ExprLStmt idL]

trS_bndrs :: [(idR, idR)]

trS_using :: LHsExpr idR

trS_by :: Maybe (LHsExpr idR)

trS_ret :: SyntaxExpr idR

trS_bind :: SyntaxExpr idR

trS_bind_arg_ty :: PostTc idR Type

trS_fmap :: HsExpr idR

RecStmt

AnnKeywordId : AnnRec

Fields

recS_stmts :: [LStmtLR idL idR body]

recS_later_ids :: [idR]

recS_rec_ids :: [idR]

recS_bind_fn :: SyntaxExpr idR

recS_ret_fn :: SyntaxExpr idR

recS_mfix_fn :: SyntaxExpr idR

recS_bind_ty :: PostTc idR Type

recS_later_rets :: [PostTcExpr]

recS_rec_rets :: [PostTcExpr]

recS_ret_ty :: PostTc idR Type

data TransForm

Constructors

ThenForm

GroupForm

data ParStmtBlock idL idR

Parenthesised Statement Block

Constructors

ParStmtBlock [ExprLStmt idL] [idR] (SyntaxExpr idR)

data ApplicativeArg idL idR

Applicative Argument

Constructors

ApplicativeArgOne (LPat idL) (LHsExpr idL)

ApplicativeArgMany [ExprLStmt idL] (HsExpr idL) (LPat idL)

pprStmt :: forall idL idR body. (OutputableBndrId idL, OutputableBndrId idR, Outputable body) => StmtLR idL idR body -> SDoc

pprTransformStmt :: OutputableBndrId id => [id] -> LHsExpr id -> Maybe (LHsExpr id) -> SDoc

pprTransStmt :: Outputable body => Maybe body -> body -> TransForm -> SDoc

pprBy :: Outputable body => Maybe body -> SDoc

pprDo :: (OutputableBndrId id, Outputable body) => HsStmtContext any -> [LStmt id body] -> SDoc

ppr_do_stmts :: (OutputableBndrId idL, OutputableBndrId idR, Outputable body) => [LStmtLR idL idR body] -> SDoc

pprComp :: (OutputableBndrId id, Outputable body) => [LStmt id body] -> SDoc

pprQuals :: (OutputableBndrId id, Outputable body) => [LStmt id body] -> SDoc

data HsSplice id

Haskell Splice

Constructors

HsTypedSplice SpliceDecoration id (LHsExpr id)

HsUntypedSplice SpliceDecoration id (LHsExpr id)

HsQuasiQuote id id SrcSpan FastString

HsSpliced ThModFinalizers (HsSplicedThing id)

data SpliceDecoration

A splice can appear with various decorations wrapped around it. This data type captures explicitly how it was originally written, for use in the pretty printer.

Constructors

HasParens
$( splice ) or $$( splice )

HasDollar
$splice or $$splice

NoParens
bare splice

isTypedSplice :: HsSplice id -> Bool

newtype ThModFinalizers

Finalizers produced by a splice with addModFinalizer

See Note [Delaying modFinalizers in untyped splices] in RnSplice. For how this is used.

Constructors

ThModFinalizers [ForeignRef (Q ())]

data HsSplicedThing id

Haskell Spliced Thing

Values that can result from running a splice.

Constructors

HsSplicedExpr (HsExpr id)
Haskell Spliced Expression

HsSplicedTy (HsType id)
Haskell Spliced Type

HsSplicedPat (Pat id)
Haskell Spliced Pattern

type SplicePointName = Name

data PendingRnSplice

Pending Renamer Splice

Constructors

PendingRnSplice UntypedSpliceFlavour SplicePointName (LHsExpr Name)

data UntypedSpliceFlavour

Constructors

UntypedExpSplice

UntypedPatSplice

UntypedTypeSplice

UntypedDeclSplice

data PendingTcSplice

Pending Type-checker Splice

Constructors

PendingTcSplice SplicePointName (LHsExpr Id)

pprPendingSplice :: OutputableBndrId id => SplicePointName -> LHsExpr id -> SDoc

pprSpliceDecl :: OutputableBndrId id => HsSplice id -> SpliceExplicitFlag -> SDoc

ppr_splice_decl :: OutputableBndrId id => HsSplice id -> SDoc

pprSplice :: OutputableBndrId id => HsSplice id -> SDoc

ppr_quasi :: OutputableBndr id => id -> id -> FastString -> SDoc

ppr_splice :: OutputableBndrId id => SDoc -> id -> LHsExpr id -> SDoc -> SDoc

data HsBracket id

Haskell Bracket

Constructors

ExpBr (LHsExpr id)

PatBr (LPat id)

DecBrL [LHsDecl id]

DecBrG (HsGroup id)

TypBr (LHsType id)

VarBr Bool id

TExpBr (LHsExpr id)

isTypedBracket :: HsBracket id -> Bool

pprHsBracket :: OutputableBndrId id => HsBracket id -> SDoc

thBrackets :: SDoc -> SDoc -> SDoc

thTyBrackets :: SDoc -> SDoc

data ArithSeqInfo id

Arithmetic Sequence Information

Constructors

From (LHsExpr id)

FromThen (LHsExpr id) (LHsExpr id)

FromTo (LHsExpr id) (LHsExpr id)

FromThenTo (LHsExpr id) (LHsExpr id) (LHsExpr id)

pp_dotdot :: SDoc

data HsMatchContext id

Haskell Match Context

Context of a pattern match. This is more subtle than it would seem. See Note [Varieties of pattern matches].

Constructors

FunRhs
A pattern matching on an argument of a function binding

Fields

mc_fun :: Located id

function binder of f

mc_fixity :: LexicalFixity

fixing of f

mc_strictness :: SrcStrictness

was the pattern banged? See Note [Varieties of binding pattern matches]

LambdaExpr
Patterns of a lambda

CaseAlt
Patterns and guards on a case alternative

IfAlt
Guards of a multi-way if alternative

ProcExpr
Patterns of a proc

PatBindRhs
A pattern binding eg [y] <- e = e

RecUpd
Record update [used only in DsExpr to tell matchWrapper what sort of runtime error message to generate]

StmtCtxt (HsStmtContext id)
Pattern of a do-stmt, list comprehension, pattern guard, etc

ThPatSplice
A Template Haskell pattern splice

ThPatQuote
A Template Haskell pattern quotation [p| (a,b) |]

PatSyn
A pattern synonym declaration

isPatSynCtxt :: HsMatchContext id -> Bool

data HsStmtContext id

Haskell Statement Context

Constructors

ListComp

MonadComp

PArrComp
Parallel array comprehension

DoExpr
do { ... }

MDoExpr
mdo { ... } ie recursive do-expression

ArrowExpr
do-notation in an arrow-command context

GhciStmtCtxt
A command-line Stmt in GHCi pat <- rhs

PatGuard (HsMatchContext id)
Pattern guard for specified thing

ParStmtCtxt (HsStmtContext id)
A branch of a parallel stmt

TransStmtCtxt (HsStmtContext id)
A branch of a transform stmt

isListCompExpr :: HsStmtContext id -> Bool

isMonadCompExpr :: HsStmtContext id -> Bool

isMonadFailStmtContext :: HsStmtContext id -> Bool

Should pattern match failure in a HsStmtContext be desugared using MonadFail?

matchSeparator :: HsMatchContext id -> SDoc

pprMatchContext :: (Outputable (NameOrRdrName id), Outputable id) => HsMatchContext id -> SDoc

pprMatchContextNoun :: (Outputable (NameOrRdrName id), Outputable id) => HsMatchContext id -> SDoc

pprAStmtContext :: (Outputable id, Outputable (NameOrRdrName id)) => HsStmtContext id -> SDoc

pprStmtContext :: (Outputable id, Outputable (NameOrRdrName id)) => HsStmtContext id -> SDoc

matchContextErrString :: Outputable id => HsMatchContext id -> SDoc

pprMatchInCtxt :: (OutputableBndrId idR, Outputable (NameOrRdrName (NameOrRdrName idR)), Outputable body) => Match idR body -> SDoc

pprStmtInCtxt :: (OutputableBndrId idL, OutputableBndrId idR, Outputable body) => HsStmtContext idL -> StmtLR idL idR body -> SDoc