HsExtension

Documentation

data GhcPass (c :: Pass)

Used as a data type index for the hsSyn AST

data Pass

Constructors

Parsed

Renamed

Typechecked

type GhcPs = GhcPass Parsed

type GhcRn = GhcPass Renamed

type GhcTc = GhcPass Typechecked

type GhcTcId = GhcTc

type family PostTc x ty

Types that are not defined until after type checking

type family PostRn x ty

Types that are not defined until after renaming

type family IdP p

Maps the "normal" id type for a given pass

type family XHsChar x

type family XHsCharPrim x

type family XHsString x

type family XHsStringPrim x

type family XHsInt x

type family XHsIntPrim x

type family XHsWordPrim x

type family XHsInt64Prim x

type family XHsWord64Prim x

type family XHsInteger x

type family XHsRat x

type family XHsFloatPrim x

type family XHsDoublePrim x

type ForallX (c :: * -> Constraint) (x :: *) = (c (XHsChar x), c (XHsCharPrim x), c (XHsString x), c (XHsStringPrim x), c (XHsInt x), c (XHsIntPrim x), c (XHsWordPrim x), c (XHsInt64Prim x), c (XHsWord64Prim x), c (XHsInteger x), c (XHsRat x), c (XHsFloatPrim x), c (XHsDoublePrim x))

Helper to apply a constraint to all extension points. It has one entry per extension point type family.

class HasSourceText a where

The SourceText fields have been moved into the extension fields, thus placing a requirement in the extension field to contain a SourceText so that the pretty printing and round tripping of source can continue to operate.

The HasSourceText class captures this requirement for the relevant fields.

Minimal complete definition

noSourceText, sourceText, setSourceText, getSourceText

Methods

noSourceText :: a

sourceText :: String -> a

setSourceText :: SourceText -> a

getSourceText :: a -> SourceText

type SourceTextX x = (HasSourceText (XHsChar x), HasSourceText (XHsCharPrim x), HasSourceText (XHsString x), HasSourceText (XHsStringPrim x), HasSourceText (XHsIntPrim x), HasSourceText (XHsWordPrim x), HasSourceText (XHsInt64Prim x), HasSourceText (XHsWord64Prim x), HasSourceText (XHsInteger x))

Provide a summary constraint that lists all the extension points requiring the HasSourceText class, so that it can be changed in one place as the named extensions change throughout the AST.

class HasDefault a where

Defaults for each annotation, used to simplify creation in arbitrary contexts

Minimal complete definition

def

Methods

def :: a

type HasDefaultX x = ForallX HasDefault x

Provide a single constraint that captures the requirement for a default across all the extension points.

class Convertable a b | a -> b where

Conversion of annotations from one type index to another. This is required where the AST is converted from one pass to another, and the extension values need to be brought along if possible. So for example a SourceText is converted via id, but needs a type signature to keep the type checker happy.

Minimal complete definition

convert

Methods

convert :: a -> b

type ConvertIdX a b = (XHsDoublePrim a ~ XHsDoublePrim b, XHsFloatPrim a ~ XHsFloatPrim b, XHsRat a ~ XHsRat b, XHsInteger a ~ XHsInteger b, XHsWord64Prim a ~ XHsWord64Prim b, XHsInt64Prim a ~ XHsInt64Prim b, XHsWordPrim a ~ XHsWordPrim b, XHsIntPrim a ~ XHsIntPrim b, XHsInt a ~ XHsInt b, XHsStringPrim a ~ XHsStringPrim b, XHsString a ~ XHsString b, XHsCharPrim a ~ XHsCharPrim b, XHsChar a ~ XHsChar b)

A constraint capturing all the extension points that can be converted via instance Convertable a a

type DataId p = (Data p, ForallX Data p, Data (NameOrRdrName (IdP p)), Data (IdP p), Data (PostRn p (IdP p)), Data (PostRn p (Located Name)), Data (PostRn p Bool), Data (PostRn p Fixity), Data (PostRn p NameSet), Data (PostRn p [Name]), Data (PostTc p (IdP p)), Data (PostTc p Coercion), Data (PostTc p ConLike), Data (PostTc p HsWrapper), Data (PostTc p Type), Data (PostTc p [ConLike]), Data (PostTc p [Type]))

type OutputableBndrId id = (OutputableBndr (NameOrRdrName (IdP id)), OutputableBndr (IdP id))

Constraint type to bundle up the requirement for OutputableBndr on both the id and the NameOrRdrName type for it