HsImpExp

Documentation

type LImportDecl name

Arguments

= Located (ImportDecl name)
When in a list this may have

AnnKeywordId : AnnSemi

Located Import Declaration

data ImportDecl name

Import Declaration

A single Haskell import declaration.

Constructors

ImportDecl
AnnKeywordIds

AnnImport

AnnOpen, AnnClose for ideclSource

AnnSafe,AnnQualified, AnnPackageName,AnnAs, AnnVal

AnnHiding,AnnOpen, AnnClose attached to location in ideclHiding

Fields

ideclSourceSrc :: SourceText

ideclName :: Located ModuleName

Module name.

ideclPkgQual :: Maybe StringLiteral

Package qualifier.

ideclSource :: Bool

True = {-# SOURCE #-} import

ideclSafe :: Bool

True => safe import

ideclQualified :: Bool

True => qualified

ideclImplicit :: Bool

True => implicit import (of Prelude)

ideclAs :: Maybe (Located ModuleName)

as Module

ideclHiding :: Maybe (Bool, Located [LIE name])

(True => hiding, names)

simpleImportDecl :: ModuleName -> ImportDecl name

data IEWrappedName name

A name in an import or export specfication which may have adornments. Used primarily for accurate pretty printing of ParsedSource, and API Annotation placement.

Constructors

IEName (Located name)
no extra

IEPattern (Located name)
pattern X

IEType (Located name)
type (:+:)

type LIEWrappedName name = Located (IEWrappedName name)

Located name with possible adornment - AnnKeywordIds : AnnType, AnnPattern

type LIE name

Arguments

= Located (IE name)
When in a list this may have

AnnKeywordId : AnnComma

Located Import or Export

data IE name

Imported or exported entity.

Constructors

IEVar (LIEWrappedName name)
Imported or Exported Variable

IEThingAbs (LIEWrappedName name)
Imported or exported Thing with Absent list

The thing is a Class/Type (can't tell) - AnnKeywordIds : AnnPattern, AnnType,AnnVal

IEThingAll (LIEWrappedName name)
Imported or exported Thing with All imported or exported

The thing is a ClassType and the All refers to methodsconstructors

AnnKeywordIds : AnnOpen, AnnDotdot,AnnClose, AnnType

IEThingWith (LIEWrappedName name) IEWildcard [LIEWrappedName name] [Located (FieldLbl name)]
Imported or exported Thing With given imported or exported

The thing is a Class/Type and the imported or exported things are methods/constructors and record fields; see Note [IEThingWith] - AnnKeywordIds : AnnOpen, AnnClose, AnnComma, AnnType

IEModuleContents (Located ModuleName)
Imported or exported module contents

(Export Only)

AnnKeywordIds : AnnModule

IEGroup Int HsDocString
Doc section heading

IEDoc HsDocString
Some documentation

IEDocNamed String
Reference to named doc

data IEWildcard

Imported or Exported Wildcard

Constructors

NoIEWildcard

IEWildcard Int

ieName :: IE name -> name

ieNames :: IE a -> [a]

ieWrappedName :: IEWrappedName name -> name

ieLWrappedName :: LIEWrappedName name -> Located name

replaceWrappedName :: IEWrappedName name1 -> name2 -> IEWrappedName name2

replaceLWrappedName :: LIEWrappedName name1 -> name2 -> LIEWrappedName name2

pprImpExp :: (HasOccName name, OutputableBndr name) => name -> SDoc