HsLit

Documentation

data HsLit

Haskell Literal

Constructors

HsChar SourceText Char
Character

HsCharPrim SourceText Char
Unboxed character

HsString SourceText FastString
String

HsStringPrim SourceText ByteString
Packed bytes

HsInt SourceText Integer
Genuinely an Int; arises from TcGenDeriv, and from TRANSLATION

HsIntPrim SourceText Integer
literal Int#

HsWordPrim SourceText Integer
literal Word#

HsInt64Prim SourceText Integer
literal Int64#

HsWord64Prim SourceText Integer
literal Word64#

HsInteger SourceText Integer Type
Genuinely an integer; arises only from TRANSLATION (overloaded literals are done with HsOverLit)

HsRat FractionalLit Type
Genuinely a rational; arises only from TRANSLATION (overloaded literals are done with HsOverLit)

HsFloatPrim FractionalLit
Unboxed Float

HsDoublePrim FractionalLit
Unboxed Double

data HsOverLit id

Haskell Overloaded Literal

Constructors

OverLit

Fields

ol_val :: OverLitVal

ol_rebindable :: PostRn id Bool

ol_witness :: HsExpr id

ol_type :: PostTc id Type

data OverLitVal

Overloaded Literal Value

Constructors

HsIntegral !SourceText !Integer
Integer-looking literals;

HsFractional !FractionalLit
Frac-looking literals

HsIsString !SourceText !FastString
String-looking literals

overLitType :: HsOverLit a -> PostTc a Type

pp_st_suffix :: SourceText -> SDoc -> SDoc -> SDoc

pmPprHsLit :: HsLit -> SDoc

pmPprHsLit pretty prints literals and is used when pretty printing pattern match warnings. All are printed the same (i.e., without hashes if they are primitive and not wrapped in constructors if they are boxed). This happens mainly for too reasons: * We do not want to expose their internal representation * The warnings become too messy