HsPat

Documentation

data Pat id

Pattern

AnnKeywordId : AnnBang

Constructors

WildPat (PostTc id Type)
Wildcard Pattern The sole reason for a type on a WildPat is to support hsPatType :: Pat Id -> Type

VarPat (Located id)
Variable Pattern

LazyPat (LPat id)
Lazy Pattern ^ - AnnKeywordId : AnnTilde

AsPat (Located id) (LPat id)
As pattern ^ - AnnKeywordId : AnnAt

ParPat (LPat id)
Parenthesised pattern See Note [Parens in HsSyn] in HsExpr ^ - AnnKeywordId : AnnOpen '(', AnnClose ')'

BangPat (LPat id)
Bang pattern ^ - AnnKeywordId : AnnBang

ListPat [LPat id] (PostTc id Type) (Maybe (PostTc id Type, SyntaxExpr id))
Syntactic List

AnnKeywordId : AnnOpen '[', AnnClose ']'

TuplePat [LPat id] Boxity [PostTc id Type]
Tuple sub-patterns

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

SumPat (LPat id) ConTag Arity (PostTc id [Type])
Anonymous sum pattern

AnnKeywordId : AnnOpen '(#', AnnClose '#)'

PArrPat [LPat id] (PostTc id Type)

AnnKeywordId : AnnOpen '[:', AnnClose ':]'

ConPatIn (Located id) (HsConPatDetails id)
Constructor Pattern In

ConPatOut
Constructor Pattern Out

Fields

pat_con :: Located ConLike

pat_arg_tys :: [Type]

pat_tvs :: [TyVar]

pat_dicts :: [EvVar]

pat_binds :: TcEvBinds

pat_args :: HsConPatDetails id

pat_wrap :: HsWrapper

ViewPat (LHsExpr id) (LPat id) (PostTc id Type)
View Pattern

SplicePat (HsSplice id)
Splice Pattern (Includes quasi-quotes)

LitPat HsLit
Literal Pattern Used for *non-overloaded* literal patterns: Int, Int, Char, String, etc.

NPat (Located (HsOverLit id)) (Maybe (SyntaxExpr id)) (SyntaxExpr id) (PostTc id Type)
Natural Pattern

AnnKeywordId : AnnVal +

NPlusKPat (Located id) (Located (HsOverLit id)) (HsOverLit id) (SyntaxExpr id) (SyntaxExpr id) (PostTc id Type)
n+k pattern

SigPatIn (LPat id) (LHsSigWcType id)
Pattern with a type signature

SigPatOut (LPat id) Type
Pattern with a type signature

CoPat HsWrapper (Pat id) Type
Coercion Pattern

type InPat id = LPat id

type OutPat id = LPat id

type LPat id = Located (Pat id)

type HsConPatDetails id = HsConDetails (LPat id) (HsRecFields id (LPat id))

Haskell Constructor Pattern Details

hsConPatArgs :: HsConPatDetails id -> [LPat id]

data HsRecFields id arg

Haskell Record Fields

HsRecFields is used only for patterns and expressions (not data type declarations)

Constructors

HsRecFields

Fields

rec_flds :: [LHsRecField id arg]

rec_dotdot :: Maybe Int

data HsRecField' id arg

Haskell Record Field

AnnKeywordId : AnnEqual,

For details on above see note [Api annotations] in ApiAnnotation

Constructors

HsRecField

Fields

hsRecFieldLbl :: Located id

hsRecFieldArg :: arg

Filled in by renamer when punning

hsRecPun :: Bool

Note [Punning]

type LHsRecField' id arg = Located (HsRecField' id arg)

Located Haskell Record Field

type HsRecField id arg = HsRecField' (FieldOcc id) arg

Haskell Record Field

type LHsRecField id arg = Located (HsRecField id arg)

Located Haskell Record Field

type HsRecUpdField id = HsRecField' (AmbiguousFieldOcc id) (LHsExpr id)

Haskell Record Update Field

type LHsRecUpdField id = Located (HsRecUpdField id)

Located Haskell Record Update Field

hsRecFields :: HsRecFields id arg -> [PostRn id id]

hsRecFieldSel :: HsRecField name arg -> Located (PostRn name name)

hsRecFieldId :: HsRecField Id arg -> Located Id

hsRecFieldsArgs :: HsRecFields id arg -> [arg]

hsRecUpdFieldId :: HsRecField' (AmbiguousFieldOcc Id) arg -> Located Id

hsRecUpdFieldOcc :: HsRecField' (AmbiguousFieldOcc Id) arg -> LFieldOcc Id

hsRecUpdFieldRdr :: HsRecUpdField id -> Located RdrName

mkPrefixConPat :: DataCon -> [OutPat id] -> [Type] -> OutPat id

mkCharLitPat :: SourceText -> Char -> OutPat id

mkNilPat :: Type -> OutPat id

looksLazyPatBind :: HsBind id -> Bool

isBangedLPat :: LPat id -> Bool

isBangedPatBind :: HsBind id -> Bool

hsPatNeedsParens :: Pat a -> Bool

isIrrefutableHsPat :: OutputableBndrId id => LPat id -> Bool

collectEvVarsPats :: [Pat id] -> Bag EvVar

pprParendLPat :: OutputableBndrId name => LPat name -> SDoc

pprConArgs :: OutputableBndrId id => HsConPatDetails id -> SDoc