HsTypes

Documentation

data HsType name

Haskell Type

Constructors

HsForAllTy

AnnKeywordId : AnnForall, AnnDot,AnnDarrow For details on above see note [Api annotations] in ApiAnnotation

Fields

hst_bndrs :: [LHsTyVarBndr name]

hst_body :: LHsType name

HsQualTy

Fields

hst_ctxt :: LHsContext name

hst_body :: LHsType name

HsTyVar Promoted (Located name)

AnnKeywordId : None

HsAppsTy [LHsAppType name]

AnnKeywordId : None

HsAppTy (LHsType name) (LHsType name)

AnnKeywordId : None

HsFunTy (LHsType name) (LHsType name)

AnnKeywordId : AnnRarrow,

HsListTy (LHsType name)

AnnKeywordId : AnnOpen '[', AnnClose ']'

HsPArrTy (LHsType name)

AnnKeywordId : AnnOpen '[:', AnnClose ':]'

HsTupleTy HsTupleSort [LHsType name]

AnnKeywordId : AnnOpen '(' or '(#', AnnClose ')' or '#)'

HsSumTy [LHsType name]

AnnKeywordId : AnnOpen '(#', AnnClose '#)'@

HsOpTy (LHsType name) (Located name) (LHsType name)

AnnKeywordId : None

HsParTy (LHsType name)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsIParamTy (Located HsIPName) (LHsType name)
(?x :: ty)

AnnKeywordId : AnnDcolon

HsEqTy (LHsType name) (LHsType name)
ty1 ~ ty2

AnnKeywordId : AnnTilde

HsKindSig (LHsType name) (LHsKind name)
(ty :: kind)

AnnKeywordId : AnnOpen '(', AnnDcolon,AnnClose ')'

HsSpliceTy (HsSplice name) (PostTc name Kind)

AnnKeywordId : AnnOpen '$(', AnnClose ')'

HsDocTy (LHsType name) LHsDocString

AnnKeywordId : None

HsBangTy HsSrcBang (LHsType name)

AnnKeywordId : AnnOpen '{-# UNPACK' or '{-# NOUNPACK', AnnClose '#-}' AnnBang '!'

HsRecTy [LConDeclField name]

AnnKeywordId : AnnOpen '{', AnnClose '}'

HsCoreTy Type

AnnKeywordId : None

HsExplicitListTy Promoted (PostTc name Kind) [LHsType name]

AnnKeywordId : AnnOpen "'[", AnnClose ']'

HsExplicitTupleTy [PostTc name Kind] [LHsType name]

AnnKeywordId : AnnOpen "'(", AnnClose ')'

HsTyLit HsTyLit

AnnKeywordId : None

HsWildCardTy (HsWildCardInfo name)

AnnKeywordId : None

type LHsType name

Arguments

= Located (HsType name)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Type

type HsKind name = HsType name

Haskell Kind

type LHsKind name

Arguments

= Located (HsKind name)
AnnKeywordId : AnnDcolon

Located Haskell Kind

data HsTyVarBndr name

Haskell Type Variable Binder

Constructors

UserTyVar (Located name)

KindedTyVar (Located name) (LHsKind name)

AnnKeywordId : AnnOpen, AnnDcolon, AnnClose

type LHsTyVarBndr name = Located (HsTyVarBndr name)

Located Haskell Type Variable Binder

data LHsQTyVars name

Located Haskell Quantified Type Variables

Constructors

HsQTvs

Fields

hsq_implicit :: PostRn name [Name]

hsq_explicit :: [LHsTyVarBndr name]

hsq_dependent :: PostRn name NameSet

data HsImplicitBndrs name thing

Haskell Implicit Binders

Constructors

HsIB

Fields

hsib_vars :: PostRn name [Name]

hsib_body :: thing

hsib_closed :: PostRn name Bool

data HsWildCardBndrs name thing

Haskell Wildcard Binders

Constructors

HsWC

Fields

hswc_wcs :: PostRn name [Name]

hswc_body :: thing

type LHsSigType name = HsImplicitBndrs name (LHsType name)

Located Haskell Signature Type

type LHsSigWcType name = HsWildCardBndrs name (LHsSigType name)

Located Haskell Signature Wildcard Type

type LHsWcType name = HsWildCardBndrs name (LHsType name)

Located Haskell Wildcard Type

data HsTupleSort

Haskell Tuple Sort

Constructors

HsUnboxedTuple

HsBoxedTuple

HsConstraintTuple

HsBoxedOrConstraintTuple

data Promoted

Promoted data types.

Constructors

Promoted

NotPromoted

type HsContext name = [LHsType name]

Haskell Context

type LHsContext name

Arguments

= Located (HsContext name)
AnnKeywordId : AnnUnit

Located Haskell Context

data HsTyLit

Haskell Type Literal

Constructors

HsNumTy SourceText Integer

HsStrTy SourceText FastString

newtype HsIPName

These names are used early on to store the names of implicit parameters. They completely disappear after type-checking.

Constructors

HsIPName FastString

hsIPNameFS :: HsIPName -> FastString

data HsAppType name

Haskell Application Type

Constructors

HsAppInfix (Located name)

HsAppPrefix (LHsType name)

type LHsAppType name

Arguments

= Located (HsAppType name)
AnnKeywordId : AnnSimpleQuote

Located Haskell Application Type

type LBangType name = Located (BangType name)

Located Bang Type

type BangType name = HsType name

Bang Type

data HsSrcBang

Haskell Source Bang

Bangs on data constructor arguments as the user wrote them in the source code.

(HsSrcBang _ SrcUnpack SrcLazy) and (HsSrcBang _ SrcUnpack NoSrcStrict) (without StrictData) makes no sense, we emit a warning (in checkValidDataCon) and treat it like (HsSrcBang _ NoSrcUnpack SrcLazy)

Constructors

HsSrcBang SourceText SrcUnpackedness SrcStrictness

data HsImplBang

Haskell Implementation Bang

Bangs of data constructor arguments as generated by the compiler after consulting HsSrcBang, flags, etc.

Constructors

HsLazy
Lazy field

HsStrict
Strict but not unpacked field

HsUnpack (Maybe Coercion)
Strict and unpacked field co :: arg-ty ~ product-ty HsBang

data SrcStrictness

Source Strictness

What strictness annotation the user wrote

Constructors

SrcLazy
Lazy, ie '~'

SrcStrict
Strict, ie !

NoSrcStrict
no strictness annotation

data SrcUnpackedness

Source Unpackedness

What unpackedness the user requested

Constructors

SrcUnpack
{--} specified

SrcNoUnpack
{--} specified

NoSrcUnpack
no unpack pragma

getBangType :: LHsType a -> LHsType a

getBangStrictness :: LHsType a -> HsSrcBang

data ConDeclField name

Constructor Declaration Field

Constructors

ConDeclField

AnnKeywordId : AnnDcolon

Fields

cd_fld_names :: [LFieldOcc name]

See Note [ConDeclField names]

cd_fld_type :: LBangType name

cd_fld_doc :: Maybe LHsDocString

type LConDeclField name

Arguments

= Located (ConDeclField name)
May have AnnKeywordId : AnnComma when in a list

Located Constructor Declaration Field

pprConDeclFields :: OutputableBndrId name => [LConDeclField name] -> SDoc

updateGadtResult

Arguments

:: Monad m

=> (SDoc -> m ())

-> SDoc

-> HsConDetails (LHsType Name) (Located [LConDeclField Name])
Original details

-> LHsType Name
Original result type

-> m (HsConDetails (LHsType Name) (Located [LConDeclField Name]), LHsType Name)

data HsConDetails arg rec

Haskell Constructor Details

Constructors

PrefixCon [arg]

RecCon rec

InfixCon arg arg

data FieldOcc name

Field Occurrence

Represents an *occurrence* of an unambiguous field. We store both the RdrName the user originally wrote, and after the renamer, the selector function.

Constructors

FieldOcc

Fields

rdrNameFieldOcc :: Located RdrName

See Note [Located RdrNames] in HsExpr

selectorFieldOcc :: PostRn name name

type LFieldOcc name = Located (FieldOcc name)

Located Field Occurrence

mkFieldOcc :: Located RdrName -> FieldOcc RdrName

data AmbiguousFieldOcc name

Ambiguous Field Occurrence

Represents an *occurrence* of a field that is potentially ambiguous after the renamer, with the ambiguity resolved by the typechecker. We always store the RdrName that the user originally wrote, and store the selector function after the renamer (for unambiguous occurrences) or the typechecker (for ambiguous occurrences).

See Note [HsRecField and HsRecUpdField] in HsPat and Note [Disambiguating record fields] in TcExpr. See Note [Located RdrNames] in HsExpr

Constructors

Unambiguous (Located RdrName) (PostRn name name)

Ambiguous (Located RdrName) (PostTc name name)

mkAmbiguousFieldOcc :: Located RdrName -> AmbiguousFieldOcc RdrName

rdrNameAmbiguousFieldOcc :: AmbiguousFieldOcc name -> RdrName

selectorAmbiguousFieldOcc :: AmbiguousFieldOcc Id -> Id

unambiguousFieldOcc :: AmbiguousFieldOcc Id -> FieldOcc Id

ambiguousFieldOcc :: FieldOcc name -> AmbiguousFieldOcc name

newtype HsWildCardInfo name

Constructors

AnonWildCard (PostRn name (Located Name))

mkAnonWildCardTy :: HsType RdrName

wildCardName :: HsWildCardInfo Name -> Name

sameWildCard :: Located (HsWildCardInfo name) -> Located (HsWildCardInfo name) -> Bool

mkHsImplicitBndrs :: thing -> HsImplicitBndrs RdrName thing

mkHsWildCardBndrs :: thing -> HsWildCardBndrs RdrName thing

hsImplicitBody :: HsImplicitBndrs name thing -> thing

mkEmptyImplicitBndrs :: thing -> HsImplicitBndrs Name thing

mkEmptyWildCardBndrs :: thing -> HsWildCardBndrs Name thing

mkHsQTvs :: [LHsTyVarBndr RdrName] -> LHsQTyVars RdrName

hsQTvExplicit :: LHsQTyVars name -> [LHsTyVarBndr name]

emptyLHsQTvs :: LHsQTyVars Name

isEmptyLHsQTvs :: LHsQTyVars Name -> Bool

isHsKindedTyVar :: HsTyVarBndr name -> Bool

Does this HsTyVarBndr come with an explicit kind annotation?

hsTvbAllKinded :: LHsQTyVars name -> Bool

Do all type variables in this LHsQTyVars come with kind annotations?

hsScopedTvs :: LHsSigType Name -> [Name]

hsWcScopedTvs :: LHsSigWcType Name -> [Name]

dropWildCards :: LHsSigWcType name -> LHsSigType name

hsTyVarName :: HsTyVarBndr name -> name

hsAllLTyVarNames :: LHsQTyVars Name -> [Name]

hsLTyVarLocNames :: LHsQTyVars name -> [Located name]

hsLTyVarName :: LHsTyVarBndr name -> name

hsLTyVarLocName :: LHsTyVarBndr name -> Located name

hsExplicitLTyVarNames :: LHsQTyVars name -> [name]

splitLHsInstDeclTy :: LHsSigType Name -> ([Name], LHsContext Name, LHsType Name)

getLHsInstDeclHead :: LHsSigType name -> LHsType name

getLHsInstDeclClass_maybe :: LHsSigType name -> Maybe (Located name)

splitLHsPatSynTy :: LHsType name -> ([LHsTyVarBndr name], LHsContext name, [LHsTyVarBndr name], LHsContext name, LHsType name)

splitLHsForAllTy :: LHsType name -> ([LHsTyVarBndr name], LHsType name)

splitLHsQualTy :: LHsType name -> (LHsContext name, LHsType name)

splitLHsSigmaTy :: LHsType name -> ([LHsTyVarBndr name], LHsContext name, LHsType name)

splitHsFunType :: LHsType Name -> ([LHsType Name], LHsType Name)

splitHsAppsTy :: [LHsAppType name] -> ([[LHsType name]], [Located name])

Splits a [HsAppType name] (the payload of an HsAppsTy) into regions of prefix types (normal types) and infix operators. If splitHsAppsTy tys = (non_syms, syms), then tys starts with the first element of non_syms followed by the first element of syms followed by the next element of non_syms, etc. It is guaranteed that the non_syms list has one more element than the syms list.

splitHsAppTys :: LHsType Name -> [LHsType Name] -> (LHsType Name, [LHsType Name])

getAppsTyHead_maybe :: [LHsAppType name] -> Maybe (LHsType name, [LHsType name], LexicalFixity)

Retrieves the head of an HsAppsTy, if this can be done unambiguously, without consulting fixities.

hsTyGetAppHead_maybe :: LHsType name -> Maybe (Located name, [LHsType name])

mkHsOpTy :: LHsType name -> Located name -> LHsType name -> HsType name

mkHsAppTy :: LHsType name -> LHsType name -> LHsType name

mkHsAppTys :: LHsType name -> [LHsType name] -> LHsType name

ignoreParens :: LHsType name -> LHsType name

hsSigType :: LHsSigType name -> LHsType name

hsSigWcType :: LHsSigWcType name -> LHsType name

hsLTyVarBndrToType :: LHsTyVarBndr name -> LHsType name

Convert a LHsTyVarBndr to an equivalent LHsType.

hsLTyVarBndrsToTypes :: LHsQTyVars name -> [LHsType name]

Convert a LHsTyVarBndrs to a list of types. Works on *type* variable only, no kind vars.

pprParendHsType :: OutputableBndrId name => HsType name -> SDoc

pprHsForAll :: OutputableBndrId name => [LHsTyVarBndr name] -> LHsContext name -> SDoc

pprHsForAllTvs :: OutputableBndrId name => [LHsTyVarBndr name] -> SDoc

pprHsForAllExtra :: OutputableBndrId name => Maybe SrcSpan -> [LHsTyVarBndr name] -> LHsContext name -> SDoc

Version of pprHsForAll that can also print an extra-constraints wildcard, e.g. _ => a -> Bool or (Show a, _) => a -> String. This underscore will be printed when the 'Maybe SrcSpan' argument is a Just containing the location of the extra-constraints wildcard. A special function for this is needed, as the extra-constraints wildcard is removed from the actual context and type, and stored in a separate field, thus just printing the type will not print the extra-constraints wildcard.

pprHsContext :: OutputableBndrId name => HsContext name -> SDoc

pprHsContextNoArrow :: OutputableBndrId name => HsContext name -> SDoc

pprHsContextMaybe :: OutputableBndrId name => HsContext name -> Maybe SDoc