IfaceSyn

Documentation

module IfaceType

data IfaceDecl

Constructors

IfaceId

Fields

ifName :: OccName

ifType :: IfaceType

ifIdDetails :: IfaceIdDetails

ifIdInfo :: IfaceIdInfo

IfaceData

Fields

ifName :: OccName

ifTyVars :: [IfaceTvBndr]

ifCtxt :: IfaceContext

ifCons :: IfaceConDecls

ifRec :: RecFlag

ifGadtSyntax :: Bool

ifFamInst :: Maybe (IfaceTyCon, [IfaceType])

IfaceSyn

Fields

ifName :: OccName

ifTyVars :: [IfaceTvBndr]

ifSynKind :: IfaceKind

ifSynRhs :: Maybe IfaceType

ifFamInst :: Maybe (IfaceTyCon, [IfaceType])

IfaceClass

Fields

ifCtxt :: IfaceContext

ifName :: OccName

ifTyVars :: [IfaceTvBndr]

ifFDs :: [FunDep FastString]

ifATs :: [IfaceAT]

ifSigs :: [IfaceClassOp]

ifRec :: RecFlag

IfaceForeign

Fields

ifName :: OccName

ifExtName :: Maybe FastString

data IfaceClassOp

Constructors

IfaceClassOp OccName DefMethSpec IfaceType

data IfaceAT

Constructors

IfaceAT IfaceDecl [IfaceATDefault]

data IfaceATDefault

Constructors

IfaceATD [IfaceTvBndr] [IfaceType] IfaceType

data IfaceConDecl

Constructors

IfCon

Fields

ifConOcc :: OccName

ifConWrapper :: Bool

ifConInfix :: Bool

ifConUnivTvs :: [IfaceTvBndr]

ifConExTvs :: [IfaceTvBndr]

ifConEqSpec :: [(OccName, IfaceType)]

ifConCtxt :: IfaceContext

ifConArgTys :: [IfaceType]

ifConFields :: [OccName]

ifConStricts :: [HsBang]

data IfaceConDecls

Constructors

IfAbstractTyCon Bool

IfOpenDataTyCon

IfDataTyCon [IfaceConDecl]

IfNewTyCon IfaceConDecl

data IfaceExpr

Constructors

IfaceLcl IfLclName

IfaceExt IfExtName

IfaceType IfaceType

IfaceCo IfaceType

IfaceTuple TupleSort [IfaceExpr]

IfaceLam IfaceBndr IfaceExpr

IfaceApp IfaceExpr IfaceExpr

IfaceCase IfaceExpr IfLclName [IfaceAlt]

IfaceLet IfaceBinding IfaceExpr

IfaceCast IfaceExpr IfaceCoercion

IfaceLit Literal

IfaceFCall ForeignCall IfaceType

IfaceTick IfaceTickish IfaceExpr

type IfaceAlt = (IfaceConAlt, [IfLclName], IfaceExpr)

data IfaceLetBndr

Constructors

IfLetBndr IfLclName IfaceType IfaceIdInfo

data IfaceBinding

Constructors

IfaceNonRec IfaceLetBndr IfaceExpr

IfaceRec [(IfaceLetBndr, IfaceExpr)]

data IfaceConAlt

Constructors

IfaceDefault

IfaceDataAlt IfExtName

IfaceLitAlt Literal

data IfaceIdInfo

Constructors

NoInfo

HasInfo [IfaceInfoItem]

data IfaceIdDetails

Constructors

IfVanillaId

IfRecSelId IfaceTyCon Bool

IfDFunId

data IfaceUnfolding

Constructors

IfCoreUnfold Bool IfaceExpr

IfCompulsory IfaceExpr

IfInlineRule Arity Bool Bool IfaceExpr

IfExtWrapper Arity IfExtName

IfLclWrapper Arity IfLclName

IfDFunUnfold [IfaceExpr]

data IfaceInfoItem

Constructors

HsArity Arity

HsStrictness StrictSig

HsInline InlinePragma

HsUnfold Bool IfaceUnfolding

HsNoCafRefs

data IfaceRule

Constructors

IfaceRule

Fields

ifRuleName :: RuleName

ifActivation :: Activation

ifRuleBndrs :: [IfaceBndr]

ifRuleHead :: IfExtName

ifRuleArgs :: [IfaceExpr]

ifRuleRhs :: IfaceExpr

ifRuleAuto :: Bool

ifRuleOrph :: Maybe OccName

data IfaceAnnotation

Constructors

IfaceAnnotation

Fields

ifAnnotatedTarget :: IfaceAnnTarget

ifAnnotatedValue :: Serialized

type IfaceAnnTarget = AnnTarget OccName

data IfaceInst

Constructors

IfaceInst

Fields

ifInstCls :: IfExtName

ifInstTys :: [Maybe IfaceTyCon]

ifDFun :: IfExtName

ifOFlag :: OverlapFlag

ifInstOrph :: Maybe OccName

data IfaceFamInst

Constructors

IfaceFamInst

Fields

ifFamInstFam :: IfExtName

ifFamInstTys :: [Maybe IfaceTyCon]

ifFamInstTyCon :: IfaceTyCon

data IfaceTickish

Constructors

IfaceHpcTick Module Int

IfaceSCC CostCentre Bool Bool

ifaceDeclSubBndrs :: IfaceDecl -> [OccName]

visibleIfConDecls :: IfaceConDecls -> [IfaceConDecl]

freeNamesIfDecl :: IfaceDecl -> NameSet

Finding the Names in IfaceSyn

freeNamesIfRule :: IfaceRule -> NameSet

pprIfaceExpr :: (SDoc -> SDoc) -> IfaceExpr -> SDoc

Pretty Print an IfaceExpre

The first argument should be a function that adds parens in context that need an atomic value (e.g. function args)

pprIfaceDeclHead :: IfaceContext -> OccName -> [IfaceTvBndr] -> SDoc