IfaceType

Documentation

type IfExtName = Name

type IfLclName = FastString

data IfaceType

Constructors

IfaceFreeTyVar TyVar

IfaceTyVar IfLclName

IfaceLitTy IfaceTyLit

IfaceAppTy IfaceType IfaceType

IfaceFunTy IfaceType IfaceType

IfaceDFunTy IfaceType IfaceType

IfaceForAllTy IfaceForAllBndr IfaceType

IfaceTyConApp IfaceTyCon IfaceTcArgs

IfaceCastTy IfaceType IfaceCoercion

IfaceCoercionTy IfaceCoercion

IfaceTupleTy TupleSort IsPromoted IfaceTcArgs

type IfacePredType = IfaceType

type IfaceKind = IfaceType

data IfaceCoercion

Constructors

IfaceReflCo Role IfaceType

IfaceFunCo Role IfaceCoercion IfaceCoercion

IfaceTyConAppCo Role IfaceTyCon [IfaceCoercion]

IfaceAppCo IfaceCoercion IfaceCoercion

IfaceForAllCo IfaceTvBndr IfaceCoercion IfaceCoercion

IfaceCoVarCo IfLclName

IfaceAxiomInstCo IfExtName BranchIndex [IfaceCoercion]

IfaceUnivCo IfaceUnivCoProv Role IfaceType IfaceType

IfaceSymCo IfaceCoercion

IfaceTransCo IfaceCoercion IfaceCoercion

IfaceNthCo Int IfaceCoercion

IfaceLRCo LeftOrRight IfaceCoercion

IfaceInstCo IfaceCoercion IfaceCoercion

IfaceCoherenceCo IfaceCoercion IfaceCoercion

IfaceKindCo IfaceCoercion

IfaceSubCo IfaceCoercion

IfaceAxiomRuleCo IfLclName [IfaceCoercion]

data IfaceUnivCoProv

Constructors

IfaceUnsafeCoerceProv

IfacePhantomProv IfaceCoercion

IfaceProofIrrelProv IfaceCoercion

IfacePluginProv String

IfaceHoleProv Unique
See Note [Holes in IfaceUnivCoProv]

data IfaceTyCon

Constructors

IfaceTyCon

Fields

ifaceTyConName :: IfExtName

ifaceTyConInfo :: IfaceTyConInfo

data IfaceTyConInfo

Constructors

IfaceTyConInfo

Fields

ifaceTyConIsPromoted :: IsPromoted

ifaceTyConSort :: IfaceTyConSort

data IfaceTyConSort

The various types of TyCons which have special, built-in syntax.

Constructors

IfaceNormalTyCon
a regular tycon

IfaceTupleTyCon !Arity !TupleSort
e.g. (a, b, c) or (). The arity is the tuple width, not the tycon arity (which is twice the width in the case of unboxed tuples).

IfaceSumTyCon !Arity
e.g. (a | b | c)

IfaceEqualityTyCon !Bool
a type equality. True indicates kind-homogeneous. See Note [Equality predicates in IfaceType] for details.

data IsPromoted

Is a TyCon a promoted data constructor or just a normal type constructor?

Constructors

IsNotPromoted

IsPromoted

data IfaceTyLit

Constructors

IfaceNumTyLit Integer

IfaceStrTyLit FastString

data IfaceTcArgs

Constructors

ITC_Nil

ITC_Vis IfaceType IfaceTcArgs

ITC_Invis IfaceKind IfaceTcArgs

type IfaceContext = [IfacePredType]

data IfaceBndr

Constructors

IfaceIdBndr !IfaceIdBndr

IfaceTvBndr !IfaceTvBndr

data IfaceOneShot

Constructors

IfaceNoOneShot

IfaceOneShot

type IfaceLamBndr = (IfaceBndr, IfaceOneShot)

type IfaceTvBndr = (IfLclName, IfaceKind)

type IfaceIdBndr = (IfLclName, IfaceType)

type IfaceTyConBinder = TyVarBndr IfaceTvBndr TyConBndrVis

type IfaceForAllBndr = TyVarBndr IfaceTvBndr ArgFlag

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [TyBinders and ArgFlags] in TyCoRep

Constructors

Required

Specified

Inferred

data ShowForAllFlag

Show forall flag

Unconditionally show the forall quantifier with (ShowForAllMust) or when (ShowForAllWhen) the names used are free in the binder or when compiling with -fprint-explicit-foralls.

Constructors

ShowForAllMust

ShowForAllWhen

ifTyConBinderTyVar :: IfaceTyConBinder -> IfaceTvBndr

Extract a IfaceTvBndr from a IfaceTyConBinder

ifTyConBinderName :: IfaceTyConBinder -> IfLclName

Extract the variable name from a IfaceTyConBinder

data IfRnEnv2

emptyIfRnEnv2 :: IfRnEnv2

eqIfaceType :: IfRnEnv2 -> IfaceType -> IfaceType -> Bool

eqIfaceTypes :: IfRnEnv2 -> [IfaceType] -> [IfaceType] -> Bool

eqIfaceTcArgs :: IfRnEnv2 -> IfaceTcArgs -> IfaceTcArgs -> Bool

eqIfaceTvBndrs :: IfRnEnv2 -> [IfaceTvBndr] -> [IfaceTvBndr] -> Maybe IfRnEnv2

Similar to eqTyVarBndrs, checks that tyvar lists are the same length and have matching kinds; if so, extend the IfRnEnv2. Returns Nothing if they don't match.

isIfaceLiftedTypeKind :: IfaceKind -> Bool

tcArgsIfaceTypes :: IfaceTcArgs -> [IfaceType]

pprIfaceType :: IfaceType -> SDoc

pprParendIfaceType :: IfaceType -> SDoc

pprIfaceContext :: [IfacePredType] -> SDoc

Prints a context or () if empty. This is used when, e.g., we want to display a context in an error message.

pprIfaceContextArr :: [IfacePredType] -> SDoc

Prints "(C a, D b) =>", including the arrow. This is used when we want to print a context in a type.

pprIfaceIdBndr :: IfaceIdBndr -> SDoc

pprIfaceLamBndr :: IfaceLamBndr -> SDoc

pprIfaceTvBndr :: Bool -> IfaceTvBndr -> SDoc

pprIfaceTyConBinders :: [IfaceTyConBinder] -> SDoc

pprIfaceBndrs :: [IfaceBndr] -> SDoc

pprIfaceTcArgs :: IfaceTcArgs -> SDoc

pprParendIfaceTcArgs :: IfaceTcArgs -> SDoc

pprIfaceForAllPart :: [IfaceForAllBndr] -> [IfacePredType] -> SDoc -> SDoc

pprIfaceForAll :: [IfaceForAllBndr] -> SDoc

Render the "forall ... ." or "forall ... ->" bit of a type.

pprIfaceSigmaType :: ShowForAllFlag -> IfaceType -> SDoc

pprIfaceTyLit :: IfaceTyLit -> SDoc

pprIfaceCoercion :: IfaceCoercion -> SDoc

pprParendIfaceCoercion :: IfaceCoercion -> SDoc

splitIfaceSigmaTy :: IfaceType -> ([IfaceForAllBndr], [IfacePredType], IfaceType)

pprIfaceTypeApp :: TyPrec -> IfaceTyCon -> IfaceTcArgs -> SDoc

pprUserIfaceForAll :: [IfaceForAllBndr] -> SDoc

pprIfaceCoTcApp :: TyPrec -> IfaceTyCon -> [IfaceCoercion] -> SDoc

pprTyTcApp :: TyPrec -> IfaceTyCon -> IfaceTcArgs -> SDoc

pprIfacePrefixApp :: TyPrec -> SDoc -> [SDoc] -> SDoc

suppressIfaceInvisibles :: DynFlags -> [IfaceTyConBinder] -> [a] -> [a]

stripIfaceInvisVars :: DynFlags -> [IfaceTyConBinder] -> [IfaceTyConBinder]

stripInvisArgs :: DynFlags -> IfaceTcArgs -> IfaceTcArgs

substIfaceType :: IfaceTySubst -> IfaceType -> IfaceType

substIfaceTyVar :: IfaceTySubst -> IfLclName -> IfaceType

substIfaceTcArgs :: IfaceTySubst -> IfaceTcArgs -> IfaceTcArgs

mkIfaceTySubst :: [IfaceTvBndr] -> [IfaceType] -> IfaceTySubst

eqIfaceTvBndr :: IfaceTvBndr -> IfaceTvBndr -> Bool