Kind

Main data type

type SuperKind = Type

"Super kinds", used to help encode Kinds as types. Invariant: a super kind is always of this form:

TyConApp SuperKindTyCon ...

type Kind = Type

The key type representing kinds in the compiler. Invariant: a kind is always in one of these forms:

FunTy k1 k2
TyConApp PrimTyCon [...]
TyVar kv   -- (during inference only)
ForAll ... -- (for top-level coercions)

typeKind :: Type -> Kind

anyKind :: Kind

See Type for details of the distinction between these Kinds

liftedTypeKind :: Kind

See Type for details of the distinction between these Kinds

unliftedTypeKind :: Kind

See Type for details of the distinction between these Kinds

openTypeKind :: Kind

See Type for details of the distinction between these Kinds

constraintKind :: Kind

See Type for details of the distinction between these Kinds

mkArrowKind :: Kind -> Kind -> Kind

Given two kinds k1 and k2, creates the Kind k1 -> k2

mkArrowKinds :: [Kind] -> Kind -> Kind

Iterated application of mkArrowKind

anyKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

liftedTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

openTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

unliftedTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

constraintKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

superKind :: Kind

See Type for details of the distinction between these Kinds

superKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

pprKind :: Kind -> SDoc

pprParendKind :: Kind -> SDoc

Deconstructing Kinds

kindAppResult :: Kind -> [Type] -> Kind

synTyConResKind :: TyCon -> Kind

Find the result Kind of a type synonym, after applying it to its arity number of type variables Actually this function works fine on data types too, but they'd always return *, so we never need to ask

splitKindFunTys :: Kind -> ([Kind], Kind)

Essentially splitFunTys on kinds

splitKindFunTysN :: Int -> Kind -> ([Kind], Kind)

Essentially splitFunTysN on kinds

splitKindFunTy_maybe :: Kind -> Maybe (Kind, Kind)

Predicates on Kinds

isLiftedTypeKind :: Kind -> Bool

isUnliftedTypeKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isOpenTypeKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isConstraintKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isConstraintOrLiftedKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

returnsConstraintKind :: Kind -> Bool

isKind :: Kind -> Bool

Is this a kind (i.e. a type-of-types)?

isKindVar :: Var -> Bool

isSuperKind :: Type -> Bool

Is this a super-kind (i.e. a type-of-kinds)?

isSuperKindTyCon :: TyCon -> Bool

isLiftedTypeKindCon :: TyCon -> Bool

isConstraintKindCon :: TyCon -> Bool

isAnyKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isAnyKindCon :: TyCon -> Bool

okArrowArgKind :: Kind -> Bool

okArrowResultKind :: Kind -> Bool

isSubOpenTypeKind :: Kind -> Bool

True of any sub-kind of OpenTypeKind

isSubOpenTypeKindKey :: Unique -> Bool

isSubKind :: Kind -> Kind -> Bool

k1 `isSubKind` k2 checks that k1 <: k2 Sub-kinding is extremely simple and does not look under arrrows or type constructors

isSubKindCon :: TyCon -> TyCon -> Bool

kc1 `isSubKindCon` kc2 checks that kc1 <: kc2

tcIsSubKind :: Kind -> Kind -> Bool

tcIsSubKindCon :: TyCon -> TyCon -> Bool

defaultKind :: Kind -> Kind

defaultKind_maybe :: Kind -> Maybe Kind

Used when generalising: default OpenKind and ArgKind to *. See Type for more information on what that means

Functions on variables

kiVarsOfKind :: Kind -> VarSet

kiVarsOfKinds :: [Kind] -> VarSet