Kind

Main data type

type Kind = Type

The key type representing kinds in the compiler.

typeKind :: Type -> Kind

Predicates on Kinds

isLiftedTypeKind :: Kind -> Bool

This version considers Constraint to be distinct from *.

isUnliftedTypeKind :: Kind -> Bool

isConstraintKind :: Kind -> Bool

returnsTyCon :: Unique -> Type -> Bool

Does the given type "end" in the given tycon? For example k -> [a] -> * ends in * and Maybe a -> [a] ends in [].

returnsConstraintKind :: Kind -> Bool

isConstraintKindCon :: TyCon -> Bool

okArrowArgKind :: Kind -> Bool

okArrowResultKind :: Kind -> Bool

classifiesTypeWithValues :: Kind -> Bool

Does this classify a type allowed to have values? Responds True to things like *, #, TYPE Lifted, TYPE v, Constraint.

True of any sub-kind of OpenTypeKind

isStarKind :: Kind -> Bool

Is this kind equivalent to *?

isStarKindSynonymTyCon :: TyCon -> Bool

Is the tycon Constraint?

isRuntimeRepPolymorphic :: Kind -> Bool

Tests whether the given type (which should look like "TYPE ...") has any free variables