Kind

Main data type

type Kind = Type

The key type representing kinds in the compiler. Invariant: a kind is always in one of these forms:

 FunTy k1 k2
 TyConApp PrimTyCon [...]
 TyVar kv   -- (during inference only)
 ForAll ... -- (for top-level coercions)

typeKind :: Type -> Kind

anyKind :: Kind

See Type for details of the distinction between these Kinds

liftedTypeKind :: Kind

See Type for details of the distinction between these Kinds

unliftedTypeKind :: Kind

See Type for details of the distinction between these Kinds

openTypeKind :: Kind

See Type for details of the distinction between these Kinds

argTypeKind :: Kind

See Type for details of the distinction between these Kinds

ubxTupleKind :: Kind

See Type for details of the distinction between these Kinds

constraintKind :: Kind

See Type for details of the distinction between these Kinds

mkArrowKind :: Kind -> Kind -> Kind

Given two kinds k1 and k2, creates the Kind k1 -> k2

mkArrowKinds :: [Kind] -> Kind -> Kind

Iterated application of mkArrowKind

anyKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

liftedTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

openTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

unliftedTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

argTypeKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

ubxTupleKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

constraintKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

tySuperKind :: SuperKind

tySuperKindTyCon :: TyCon

See Type for details of the distinction between the Kind TyCons

pprKind :: Kind -> SDoc

pprParendKind :: Kind -> SDoc

Deconstructing Kinds

kindFunResult :: Kind -> KindOrType -> Kind

Essentially funResultTy on kinds handling pi-types too

kindAppResult :: Kind -> [Type] -> Kind

synTyConResKind :: TyCon -> Kind

Find the result Kind of a type synonym, after applying it to its arity number of type variables Actually this function works fine on data types too, but they'd always return *, so we never need to ask

splitKindFunTys :: Kind -> ([Kind], Kind)

Essentially splitFunTys on kinds

splitKindFunTysN :: Int -> Kind -> ([Kind], Kind)

Essentially splitFunTysN on kinds

splitKindFunTy_maybe :: Kind -> Maybe (Kind, Kind)

Predicates on Kinds

isLiftedTypeKind :: Kind -> Bool

isUnliftedTypeKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isOpenTypeKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isUbxTupleKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isArgTypeKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isConstraintKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isConstraintOrLiftedKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isKind :: Kind -> Bool

Is this a kind (i.e. a type-of-types)?

isSuperKind :: Type -> Bool

Is this a super-kind (i.e. a type-of-kinds)?

noHashInKind :: Kind -> Bool

isLiftedTypeKindCon :: TyCon -> Bool

isConstraintKindCon :: TyCon -> Bool

isAnyKind :: Kind -> Bool

See Type for details of the distinction between these Kinds

isAnyKindCon :: TyCon -> Bool

isSubArgTypeKind :: Kind -> Bool

True of any sub-kind of ArgTypeKind

tcIsSubArgTypeKind :: Kind -> Bool

True of any sub-kind of ArgTypeKind

isSubOpenTypeKind :: Kind -> Bool

True of any sub-kind of OpenTypeKind

True of any sub-kind of OpenTypeKind

tcIsSubOpenTypeKind :: Kind -> Bool

True of any sub-kind of OpenTypeKind

isSubKind :: Kind -> Kind -> Bool

tcIsSubKind :: Kind -> Kind -> Bool

defaultKind :: Kind -> Kind

Used when generalising: default OpenKind and ArgKind to *. See Type for more information on what that means

isSubKindCon :: TyCon -> TyCon -> Bool

kc1 `isSubKindCon` kc2 checks that kc1 <: kc2

tcIsSubKindCon :: TyCon -> TyCon -> Bool

isSubOpenTypeKindCon :: TyCon -> Bool

Functions on variables

isKiVar :: TyVar -> Bool

splitKiTyVars :: [TyVar] -> ([KindVar], [TyVar])

partitionKiTyVars :: [TyVar] -> ([KindVar], [TyVar])

kiVarsOfKind :: Kind -> VarSet

kiVarsOfKinds :: [Kind] -> VarSet

Promotion related functions

promoteType :: Type -> Kind

Promotes a type to a kind. Assumes the argument is promotable.

isPromotableType :: Type -> Bool

isPromotableKind :: Kind -> Maybe Int