KnownUniques

Description

This is where we define a mapping from Uniques to their associated known-key Names for things associated with tuples and sums. We use this mapping while deserializing known-key Names in interface file symbol tables, which are encoded as their Unique. See Note [Symbol table representation of names] for details.

Looking up known-key names

knownUniqueName :: Unique -> Maybe Name

Get the Name associated with a known-key Unique.

Getting the Uniques of Names

Anonymous sums

mkSumTyConUnique :: Arity -> Unique

mkSumDataConUnique :: ConTagZ -> Arity -> Unique

Tuples

Vanilla

mkTupleTyConUnique :: Boxity -> Arity -> Unique

mkTupleDataConUnique :: Boxity -> Arity -> Unique

Constraint

mkCTupleTyConUnique :: Arity -> Unique

mkCTupleDataConUnique :: Arity -> Unique