Language.Haskell.Syntax.Expr

Description

Abstract Haskell syntax for expressions.

Documentation

newtype FieldLabelStrings p

RecordDotSyntax field updates

Constructors

FieldLabelStrings [Located (HsFieldLabel p)]

pprFieldLabelStrings :: FieldLabelStrings p -> SDoc

type RecProj p arg = HsRecField' (FieldLabelStrings p) arg

type LHsRecProj p arg = XRec p (RecProj p arg)

type RecUpdProj p = RecProj p (LHsExpr p)

type LHsRecUpdProj p = XRec p (RecUpdProj p)

Expressions proper

type LHsExpr p

Arguments

= XRec p (HsExpr p)
May have AnnKeywordId : AnnComma when in a list

Located Haskell Expression

type family SyntaxExpr p

Syntax Expression

SyntaxExpr is represents the function used in interpreting rebindable syntax. In the parser, we have no information to supply; in the renamer, we have the name of the function (but see Note [Monad fail : Rebindable syntax, overloaded strings] for a wrinkle) and in the type-checker we have a more elaborate structure SyntaxExprTc.

In some contexts, rebindable syntax is not implemented, and so we have constructors to represent that possibility in both the renamer and typechecker instantiations.

E.g. (>>=) is filled in before the renamer by the appropriate Name for (>>=), and then instantiated by the type checker with its type args etc

type CmdSyntaxTable p = [(Name, HsExpr p)]

Command Syntax Table (for Arrow syntax)

data HsExpr p

A Haskell expression.

Constructors

HsVar (XVar p) (LIdP p)
Variable See Note [Located RdrNames]

HsUnboundVar (XUnboundVar p) OccName
Unbound variable; also used for "holes" (_ or _x). Turned from HsVar to HsUnboundVar by the renamer, when it finds an out-of-scope variable or hole. The (XUnboundVar p) field becomes an HoleExprRef after typechecking; this is where the erroring expression will be written after solving. See Note [Holes] in GHC.Tc.Types.Constraint.

HsConLikeOut (XConLikeOut p) ConLike
After typechecker only; must be different HsVar for pretty printing

HsRecFld (XRecFld p) (AmbiguousFieldOcc p)
Variable pointing to record selector The parser produces HsVars The renamer renames record-field selectors to HsRecFld The typechecker preserves HsRecFld

HsOverLabel (XOverLabel p) FastString
Overloaded label (Note [Overloaded labels] in GHC.OverloadedLabels)

HsIPVar (XIPVar p) HsIPName
Implicit parameter (not in use after typechecking)

HsOverLit (XOverLitE p) (HsOverLit p)
Overloaded literals

HsLit (XLitE p) (HsLit p)
Simple (non-overloaded) literals

HsLam (XLam p) (MatchGroup p (LHsExpr p))
Lambda abstraction. Currently always a single match

AnnKeywordId : AnnLam, AnnRarrow,

HsLamCase (XLamCase p) (MatchGroup p (LHsExpr p))
Lambda-case

AnnKeywordId : AnnLam, AnnCase,AnnOpen, AnnClose

HsApp (XApp p) (LHsExpr p) (LHsExpr p)
Application

HsAppType (XAppTypeE p) (LHsExpr p) (LHsWcType (NoGhcTc p))
Visible type application

Explicit type argument; e.g f @Int x y NB: Has wildcards, but no implicit quantification

AnnKeywordId : AnnAt,

OpApp (XOpApp p) (LHsExpr p) (LHsExpr p) (LHsExpr p)
Operator applications: NB Bracketed ops such as (+) come out as Vars.

NegApp (XNegApp p) (LHsExpr p) (SyntaxExpr p)
Negation operator. Contains the negated expression and the name of negate

AnnKeywordId : AnnMinus

HsPar

AnnKeywordId : AnnOpen '(', AnnClose ')'

Fields

(XPar p)

(LHsExpr p)

Parenthesised expr; see Note [Parens in HsSyn]

SectionL (XSectionL p) (LHsExpr p) (LHsExpr p)

SectionR (XSectionR p) (LHsExpr p) (LHsExpr p)

ExplicitTuple (XExplicitTuple p) [HsTupArg p] Boxity
Used for explicit tuples and sections thereof

AnnKeywordId : AnnOpen, AnnClose

ExplicitSum (XExplicitSum p) ConTag Arity (LHsExpr p)
Used for unboxed sum types

AnnKeywordId : AnnOpen '(#', AnnVbar, AnnClose '#)',

There will be multiple AnnVbar, (1 - alternative) before the expression, (arity - alternative) after it

HsCase (XCase p) (LHsExpr p) (MatchGroup p (LHsExpr p))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsIf (XIf p) (LHsExpr p) (LHsExpr p) (LHsExpr p)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsMultiIf (XMultiIf p) [LGRHS p (LHsExpr p)]
Multi-way if

AnnKeywordId : AnnIf AnnOpen,AnnClose,

HsLet (XLet p) (HsLocalBinds p) (LHsExpr p)
let(rec)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsDo (XDo p) (HsStmtContext (HsDoRn p)) (XRec p [ExprLStmt p])

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

ExplicitList (XExplicitList p) [LHsExpr p]
Syntactic list: [a,b,c,...]

AnnKeywordId : AnnOpen '[', AnnClose ']'

RecordCon
Record construction

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}'

Fields

rcon_ext :: XRecordCon p

rcon_con :: XRec p (ConLikeP p)

rcon_flds :: HsRecordBinds p

RecordUpd
Record update

AnnKeywordId : AnnOpen '{', AnnDotdot,AnnClose '}' 'GHC.Parser.Annotation.AnnComma, AnnDot, AnnClose '}'

Fields

rupd_ext :: XRecordUpd p

rupd_expr :: LHsExpr p

rupd_flds :: Either [LHsRecUpdField p] [LHsRecUpdProj p]

HsGetField
Record field selection e.g z.x.

AnnKeywordId : AnnDot

This case only arises when the OverloadedRecordDot langauge extension is enabled.

Fields

gf_ext :: XGetField p

gf_expr :: LHsExpr p

gf_field :: Located (HsFieldLabel p)

HsProjection
Record field selector. e.g. (.x) or (.x.y)

AnnKeywordId : AnnOpenP AnnDot, AnnCloseP

This case only arises when the OverloadedRecordDot langauge extensions is enabled.

Fields

proj_ext :: XProjection p

proj_flds :: NonEmpty (Located (HsFieldLabel p))

ExprWithTySig (XExprWithTySig p) (LHsExpr p) (LHsSigWcType (NoGhcTc p))
Expression with an explicit type signature. e :: type

AnnKeywordId : AnnDcolon

ArithSeq (XArithSeq p) (Maybe (SyntaxExpr p)) (ArithSeqInfo p)
Arithmetic sequence

AnnKeywordId : AnnOpen '[', AnnComma,AnnDotdot, AnnClose ']'

HsBracket (XBracket p) (HsBracket p)

AnnKeywordId : AnnOpen, AnnOpenE,AnnOpenEQ, AnnClose,AnnCloseQ

HsRnBracketOut (XRnBracketOut p) (HsBracket (HsBracketRn p)) [PendingRnSplice' p]

HsTcBracketOut (XTcBracketOut p) (Maybe QuoteWrapper) (HsBracket (HsBracketRn p)) [PendingTcSplice' p]

HsSpliceE (XSpliceE p) (HsSplice p)

AnnKeywordId : AnnOpen, AnnClose

HsProc (XProc p) (LPat p) (LHsCmdTop p)
proc notation for Arrows

AnnKeywordId : AnnProc, AnnRarrow

HsStatic (XStatic p) (LHsExpr p)

AnnKeywordId : AnnStatic,

HsTick (XTick p) CoreTickish (LHsExpr p)

HsBinTick (XBinTick p) Int Int (LHsExpr p)

HsPragE (XPragE p) (HsPragE p) (LHsExpr p)

XExpr !(XXExpr p)

type family HsDoRn p

The AST used to hard-refer to GhcPass, which was a layer violation. For now, we paper it over with this new extension point.

type family HsBracketRn p

type family PendingRnSplice' p

type family PendingTcSplice' p

data HsFieldLabel p

Constructors

HsFieldLabel

Fields

hflExt :: XCHsFieldLabel p

hflLabel :: Located FieldLabelString

XHsFieldLabel !(XXHsFieldLabel p)

data HsPragE p

A pragma, written as {-# ... #-}, that may appear within an expression.

Constructors

HsPragSCC (XSCC p) SourceText StringLiteral

XHsPragE !(XXPragE p)

AnnKeywordId : AnnOpen, AnnOpen '{-# GENERATED', AnnVal,AnnVal, AnnColon,AnnVal, AnnMinus, AnnVal,AnnColon, AnnVal, AnnClose '#-}'

type LHsTupArg id = XRec id (HsTupArg id)

Located Haskell Tuple Argument

HsTupArg is used for tuple sections (,a,) is represented by ExplicitTuple [Missing ty1, Present a, Missing ty3] Which in turn stands for (x:ty1 y:ty2. (x,a,y))

data HsTupArg id

AnnKeywordId : AnnComma

Haskell Tuple Argument

Constructors

Present (XPresent id) (LHsExpr id)
The argument

Missing (XMissing id)
The argument is missing, but this is its type

XTupArg !(XXTupArg id)
Note [Trees that Grow] extension point

pprExternalSrcLoc :: (StringLiteral, (Int, Int), (Int, Int)) -> SDoc

type LHsCmd id = XRec id (HsCmd id)

Located Haskell Command (for arrow syntax)

data HsCmd id

Haskell Command (e.g. a "statement" in an Arrow proc block)

Constructors

HsCmdArrApp (XCmdArrApp id) (LHsExpr id) (LHsExpr id) HsArrAppType Bool

AnnKeywordId : Annlarrowtail, Annrarrowtail,AnnLarrowtail, AnnRarrowtail

HsCmdArrForm (XCmdArrForm id) (LHsExpr id) LexicalFixity (Maybe Fixity) [LHsCmdTop id]

AnnKeywordId : AnnOpenB '(|', AnnCloseB '|)'

HsCmdApp (XCmdApp id) (LHsCmd id) (LHsExpr id)

HsCmdLam (XCmdLam id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnLam, AnnRarrow,

HsCmdPar (XCmdPar id) (LHsCmd id)

AnnKeywordId : AnnOpen '(', AnnClose ')'

HsCmdCase (XCmdCase id) (LHsExpr id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnCase, AnnOf,AnnOpen '{', AnnClose '}'

HsCmdLamCase (XCmdLamCase id) (MatchGroup id (LHsCmd id))

AnnKeywordId : AnnLam, AnnCase,AnnOpen '{', AnnClose '}'

HsCmdIf (XCmdIf id) (SyntaxExpr id) (LHsExpr id) (LHsCmd id) (LHsCmd id)

AnnKeywordId : AnnIf, AnnSemi, AnnThen,AnnSemi, AnnElse,

HsCmdLet (XCmdLet id) (HsLocalBinds id) (LHsCmd id)

AnnKeywordId : AnnLet, AnnOpen '{', AnnClose '}',AnnIn

HsCmdDo (XCmdDo id) (XRec id [CmdLStmt id])

AnnKeywordId : AnnDo, AnnOpen, AnnSemi, AnnVbar, AnnClose

XCmd !(XXCmd id)

data HsArrAppType

Haskell Array Application Type

Constructors

HsHigherOrderApp

HsFirstOrderApp

type LHsCmdTop p = XRec p (HsCmdTop p)

Top-level command, introducing a new arrow. This may occur inside a proc (where the stack is empty) or as an argument of a command-forming operator.

Located Haskell Top-level Command

data HsCmdTop p

Haskell Top-level Command

Constructors

HsCmdTop (XCmdTop p) (LHsCmd p)

XCmdTop !(XXCmdTop p)

type HsRecordBinds p = HsRecFields p (LHsExpr p)

Haskell Record Bindings

data MatchGroup p body

Constructors

MG

Fields

mg_ext :: XMG p body

mg_alts :: XRec p [LMatch p body]

mg_origin :: Origin

XMatchGroup !(XXMatchGroup p body)

data MatchGroupTc

Constructors

MatchGroupTc

Fields

mg_arg_tys :: [Scaled Type]

mg_res_ty :: Type

type LMatch id body = XRec id (Match id body)

Located Match

May have AnnKeywordId : AnnSemi when in a list

data Match p body

Constructors

Match

Fields

m_ext :: XCMatch p body

m_ctxt :: HsMatchContext (NoGhcTc p)

m_pats :: [LPat p]

m_grhss :: GRHSs p body

XMatch !(XXMatch p body)

isInfixMatch :: Match id body -> Bool

data GRHSs p body

Guarded Right-Hand Sides

GRHSs are used both for pattern bindings and for Matches

AnnKeywordId : AnnVbar, AnnEqual,AnnWhere, AnnOpen,AnnClose AnnRarrow,AnnSemi

Constructors

GRHSs

Fields

grhssExt :: XCGRHSs p body

grhssGRHSs :: [LGRHS p body]

Guarded RHSs

grhssLocalBinds :: HsLocalBinds p

The where clause

XGRHSs !(XXGRHSs p body)

type LGRHS id body = XRec id (GRHS id body)

Located Guarded Right-Hand Side

data GRHS p body

Guarded Right Hand Side.

Constructors

GRHS (XCGRHS p body) [GuardLStmt p] body

XGRHS !(XXGRHS p body)

type LStmt id body = XRec id (StmtLR id id body)

Located do block Statement

type LStmtLR idL idR body = XRec idL (StmtLR idL idR body)

Located Statement with separate Left and Right id's

type Stmt id body = StmtLR id id body

do block Statement

type CmdLStmt id = LStmt id (LHsCmd id)

Command Located Statement

type CmdStmt id = Stmt id (LHsCmd id)

Command Statement

type ExprLStmt id = LStmt id (LHsExpr id)

Expression Located Statement

type ExprStmt id = Stmt id (LHsExpr id)

Expression Statement

type GuardLStmt id = LStmt id (LHsExpr id)

Guard Located Statement

type GuardStmt id = Stmt id (LHsExpr id)

Guard Statement

type GhciLStmt id = LStmt id (LHsExpr id)

Ghci Located Statement

type GhciStmt id = Stmt id (LHsExpr id)

Ghci Statement

data StmtLR idL idR body

Exact print annotations when in qualifier lists or guards - AnnKeywordId : AnnVbar, AnnComma,AnnThen, AnnBy,AnnBy, AnnGroup,AnnUsing

Constructors

LastStmt (XLastStmt idL idR body) body (Maybe Bool) (SyntaxExpr idR)

BindStmt

Fields

(XBindStmt idL idR body)

Post renaming has optional fail and bind / (>>=) operator. Post typechecking, also has multiplicity of the argument and the result type of the function passed to bind; that is, (P, S) in (>>=) :: Q -> (R # P -> S) -> T See Note [The type of bind in Stmts]

(LPat idL)

body

ApplicativeStmt (XApplicativeStmt idL idR body) [(SyntaxExpr idR, ApplicativeArg idL)] (Maybe (SyntaxExpr idR))
ApplicativeStmt represents an applicative expression built with <$> and <*>. It is generated by the renamer, and is desugared into the appropriate applicative expression by the desugarer, but it is intended to be invisible in error messages.

For full details, see Note [ApplicativeDo] in GHC.Rename.Expr

BodyStmt (XBodyStmt idL idR body) body (SyntaxExpr idR) (SyntaxExpr idR)

LetStmt (XLetStmt idL idR body) (HsLocalBindsLR idL idR)

AnnKeywordId : AnnLet AnnOpen '{',AnnClose '}',

ParStmt (XParStmt idL idR body) [ParStmtBlock idL idR] (HsExpr idR) (SyntaxExpr idR)

TransStmt

Fields

trS_ext :: XTransStmt idL idR body

trS_form :: TransForm

trS_stmts :: [ExprLStmt idL]

trS_bndrs :: [(IdP idR, IdP idR)]

trS_using :: LHsExpr idR

trS_by :: Maybe (LHsExpr idR)

trS_ret :: SyntaxExpr idR

trS_bind :: SyntaxExpr idR

trS_fmap :: HsExpr idR

RecStmt

AnnKeywordId : AnnRec

Fields

recS_ext :: XRecStmt idL idR body

recS_stmts :: XRec idR [LStmtLR idL idR body]

recS_later_ids :: [IdP idR]

recS_rec_ids :: [IdP idR]

recS_bind_fn :: SyntaxExpr idR

recS_ret_fn :: SyntaxExpr idR

recS_mfix_fn :: SyntaxExpr idR

XStmtLR !(XXStmtLR idL idR body)

data TransForm

Constructors

ThenForm

GroupForm

data ParStmtBlock idL idR

Parenthesised Statement Block

Constructors

ParStmtBlock (XParStmtBlock idL idR) [ExprLStmt idL] [IdP idR] (SyntaxExpr idR)

XParStmtBlock !(XXParStmtBlock idL idR)

type FailOperator id = Maybe (SyntaxExpr id)

The fail operator

This is used for `.. <-` "bind statements" in do notation, including non-monadic "binds" in applicative.

The fail operator is 'Just expr' if it potentially fail monadically. if the pattern match cannot fail, or shouldn't fail monadically (regular incomplete pattern exception), it is Nothing.

See Note [Monad fail : Rebindable syntax, overloaded strings] for the type of expression in the Just case, and why it is so.

See Note [Failing pattern matches in Stmts] for which contexts for 'BindStmt's should use the monadic fail and which shouldn't.

data ApplicativeArg idL

Applicative Argument

Constructors

ApplicativeArgOne

Fields

xarg_app_arg_one :: XApplicativeArgOne idL

The fail operator, after renaming

The fail operator is needed if this is a BindStmt where the pattern can fail. E.g.: (Just a) <- stmt The fail operator will be invoked if the pattern match fails. It is also used for guards in MonadComprehensions. The fail operator is Nothing if the pattern match can't fail

app_arg_pattern :: LPat idL

arg_expr :: LHsExpr idL

is_body_stmt :: Bool

True = was a BodyStmt, False = was a BindStmt. See Note [Applicative BodyStmt]

ApplicativeArgMany

Fields

xarg_app_arg_many :: XApplicativeArgMany idL

app_stmts :: [ExprLStmt idL]

final_expr :: HsExpr idL

bv_pattern :: LPat idL

stmt_context :: HsStmtContext (ApplicativeArgStmCtxPass idL)

context of the do expression, used in pprArg

XApplicativeArg !(XXApplicativeArg idL)

type family ApplicativeArgStmCtxPass idL

data HsSplice id

Haskell Splice

Constructors

HsTypedSplice (XTypedSplice id) SpliceDecoration (IdP id) (LHsExpr id)

HsUntypedSplice (XUntypedSplice id) SpliceDecoration (IdP id) (LHsExpr id)

HsQuasiQuote (XQuasiQuote id) (IdP id) (IdP id) SrcSpan FastString

HsSpliced (XSpliced id) ThModFinalizers (HsSplicedThing id)

XSplice !(XXSplice id)

data SpliceDecoration

A splice can appear with various decorations wrapped around it. This data type captures explicitly how it was originally written, for use in the pretty printer.

Constructors

DollarSplice
$splice or $$splice

BareSplice
bare splice

isTypedSplice :: HsSplice id -> Bool

newtype ThModFinalizers

Finalizers produced by a splice with addModFinalizer

See Note [Delaying modFinalizers in untyped splices] in GHC.Rename.Splice. For how this is used.

Constructors

ThModFinalizers [ForeignRef (Q ())]

data HsSplicedThing id

Haskell Spliced Thing

Values that can result from running a splice.

Constructors

HsSplicedExpr (HsExpr id)
Haskell Spliced Expression

HsSplicedTy (HsType id)
Haskell Spliced Type

HsSplicedPat (Pat id)
Haskell Spliced Pattern

type SplicePointName = Name

data UntypedSpliceFlavour

Constructors

UntypedExpSplice

UntypedPatSplice

UntypedTypeSplice

UntypedDeclSplice

data HsBracket p

Haskell Bracket

Constructors

ExpBr (XExpBr p) (LHsExpr p)

PatBr (XPatBr p) (LPat p)

DecBrL (XDecBrL p) [LHsDecl p]

DecBrG (XDecBrG p) (HsGroup p)

TypBr (XTypBr p) (LHsType p)

VarBr (XVarBr p) Bool (LIdP p)

TExpBr (XTExpBr p) (LHsExpr p)

XBracket !(XXBracket p)

isTypedBracket :: HsBracket id -> Bool

data ArithSeqInfo id

Arithmetic Sequence Information

Constructors

From (LHsExpr id)

FromThen (LHsExpr id) (LHsExpr id)

FromTo (LHsExpr id) (LHsExpr id)

FromThenTo (LHsExpr id) (LHsExpr id) (LHsExpr id)

data HsMatchContext p

Haskell Match Context

Context of a pattern match. This is more subtle than it would seem. See Note [Varieties of pattern matches].

Constructors

FunRhs

Fields

mc_fun :: LIdP p

function binder of f

mc_fixity :: LexicalFixity

fixing of f

mc_strictness :: SrcStrictness

was f banged? See Note [FunBind vs PatBind]

LambdaExpr
Patterns of a lambda

CaseAlt
Patterns and guards on a case alternative

IfAlt
Guards of a multi-way if alternative

ArrowMatchCtxt HsArrowMatchContext
A pattern match inside arrow notation

PatBindRhs
A pattern binding eg [y] <- e = e

PatBindGuards
Guards of pattern bindings, e.g., (Just b) | Just _ <- x = e | otherwise = e'

RecUpd
Record update [used only in GHC.HsToCore.Expr to tell matchWrapper what sort of runtime error message to generate]

StmtCtxt (HsStmtContext p)
Pattern of a do-stmt, list comprehension, pattern guard, etc

ThPatSplice
A Template Haskell pattern splice

ThPatQuote
A Template Haskell pattern quotation [p| (a,b) |]

PatSyn
A pattern synonym declaration

isPatSynCtxt :: HsMatchContext p -> Bool

data HsStmtContext p

Haskell Statement Context.

Constructors

ListComp

MonadComp

DoExpr (Maybe ModuleName)

ModuleName.

do { ... }

MDoExpr (Maybe ModuleName)

ModuleName.

mdo { ... } ie recursive do-expression

ArrowExpr
do-notation in an arrow-command context

GhciStmtCtxt
A command-line Stmt in GHCi pat <- rhs

PatGuard (HsMatchContext p)
Pattern guard for specified thing

ParStmtCtxt (HsStmtContext p)
A branch of a parallel stmt

TransStmtCtxt (HsStmtContext p)
A branch of a transform stmt

data HsArrowMatchContext

Haskell arrow match context.

Constructors

ProcExpr
A proc expression

ArrowCaseAlt
A case alternative inside arrow notation

KappaExpr
An arrow kappa abstraction

qualifiedDoModuleName_maybe :: HsStmtContext p -> Maybe ModuleName

isComprehensionContext :: HsStmtContext id -> Bool

isMonadStmtContext :: HsStmtContext id -> Bool

Is this a monadic context?

isMonadCompContext :: HsStmtContext id -> Bool

matchSeparator :: HsMatchContext p -> SDoc

pprMatchContext :: (Outputable (IdP p), UnXRec p) => HsMatchContext p -> SDoc

pprMatchContextNoun :: forall p. (Outputable (IdP p), UnXRec p) => HsMatchContext p -> SDoc

pprArrowMatchContextNoun :: HsArrowMatchContext -> SDoc

pprAStmtContext :: (Outputable (IdP p), UnXRec p) => HsStmtContext p -> SDoc

pprStmtContext :: (Outputable (IdP p), UnXRec p) => HsStmtContext p -> SDoc

prependQualified :: Maybe ModuleName -> SDoc -> SDoc