Language.Haskell.Syntax.ImpExp

Documentation

type LImportDecl pass

Arguments

= XRec pass (ImportDecl pass)
When in a list this may have

AnnKeywordId : AnnSemi

Located Import Declaration

data ImportDeclQualifiedStyle

If/how an import is qualified.

Constructors

QualifiedPre
qualified appears in prepositive position.

QualifiedPost
qualified appears in postpositive position.

NotQualified
Not qualified.

data IsBootInterface

Indicates whether a module name is referring to a boot interface (hs-boot file) or regular module (hs file). We need to treat boot modules specially when building compilation graphs, since they break cycles. Regular source files and signature files are treated equivalently.

Constructors

NotBoot

IsBoot

data ImportDecl pass

Import Declaration

A single Haskell import declaration.

Constructors

ImportDecl

Fields

ideclExt :: XCImportDecl pass

ideclName :: XRec pass ModuleName

Module name.

ideclPkgQual :: ImportDeclPkgQual pass

Package qualifier.

ideclSource :: IsBootInterface

IsBoot = {-# SOURCE #-} import

ideclSafe :: Bool

True => safe import

ideclQualified :: ImportDeclQualifiedStyle

If/how the import is qualified.

ideclAs :: Maybe (XRec pass ModuleName)

as Module

ideclImportList :: Maybe (ImportListInterpretation, XRec pass [LIE pass])

Explicit import list (EverythingBut => hiding, names)

XImportDecl !(XXImportDecl pass)
AnnKeywordIds

AnnImport

AnnOpen, AnnClose for ideclSource

AnnSafe,AnnQualified, AnnPackageName,AnnAs, AnnVal

AnnHiding,AnnOpen, AnnClose attached to location in ideclImportList

data ImportListInterpretation

Whether the import list is exactly what to import, or whether hiding was used, and therefore everything but what was listed should be imported

Constructors

Exactly

EverythingBut

type LIE pass

Arguments

= XRec pass (IE pass)
When in a list this may have

AnnKeywordId : AnnComma

Located Import or Export

data IE pass

Imported or exported entity.

Constructors

IEVar (XIEVar pass) (LIEWrappedName pass)
Imported or Exported Variable

IEThingAbs (XIEThingAbs pass) (LIEWrappedName pass)
Imported or exported Thing with Absent list

The thing is a Class/Type (can't tell) - AnnKeywordIds : AnnPattern, AnnType,AnnVal

IEThingAll (XIEThingAll pass) (LIEWrappedName pass)
Imported or exported Thing with All imported or exported

The thing is a ClassType and the All refers to methodsconstructors

AnnKeywordIds : AnnOpen, AnnDotdot,AnnClose, AnnType

IEThingWith (XIEThingWith pass) (LIEWrappedName pass) IEWildcard [LIEWrappedName pass]
Imported or exported Thing With given imported or exported

The thing is a Class/Type and the imported or exported things are methods/constructors and record fields; see Note [IEThingWith] - AnnKeywordIds : AnnOpen, AnnClose, AnnComma, AnnType

IEModuleContents (XIEModuleContents pass) (XRec pass ModuleName)
Imported or exported module contents

(Export Only)

AnnKeywordIds : AnnModule

IEGroup (XIEGroup pass) Int (LHsDoc pass)
Doc section heading

IEDoc (XIEDoc pass) (LHsDoc pass)
Some documentation

IEDocNamed (XIEDocNamed pass) String
Reference to named doc

XIE !(XXIE pass)

data IEWildcard

Wildcard in an import or export sublist, like the .. in import Mod ( T(Mk1, Mk2, ..) ).

Constructors

NoIEWildcard
no wildcard in this list

IEWildcard Int
wildcard after the given # of items in this list The Int is in the range [0..n], where n is the length of the list.

data IEWrappedName p

A name in an import or export specification which may have adornments. Used primarily for accurate pretty printing of ParsedSource, and API Annotation placement. The Annotation is the location of the adornment in the original source.

Constructors

IEName (XIEName p) (LIdP p)
no extra

IEPattern (XIEPattern p) (LIdP p)
pattern X

IEType (XIEType p) (LIdP p)
type (:+:)

XIEWrappedName !(XXIEWrappedName p)

type LIEWrappedName p = XRec p (IEWrappedName p)

Located name with possible adornment - AnnKeywordIds : AnnType, AnnPattern