Llvm.MetaData

Metadata can be constructed in a few different ways (See below).

After which it can either be attached to LLVM statements to pass along

Metadata expressions -- these are the raw metadata values that encode

Metadata nodes -- global metadata variables that attach a metadata

Named metadata -- global metadata variables that attach a metadata nodes

Attach to instructions -- metadata can be attached to LLVM instructions

As arguments -- llvm functions can take metadata as arguments, for

As a named metadata -- Here the metadata is simply declared in global

newtype MetaId

A reference to an un-named metadata node.

Constructors

MetaId Int

data MetaExpr

LLVM metadata expressions

Constructors

MetaStr !LMString

MetaNode !MetaId

MetaVar !LlvmVar

MetaStruct [MetaExpr]

data MetaAnnot

Associates some metadata with a specific label for attaching to an instruction.

Constructors

MetaAnnot LMString MetaExpr

data MetaDecl

Metadata declarations. Metadata can only be declared in global scope.

Constructors

MetaNamed !LMString [MetaId]
Named metadata. Only used for communicating module information to LLVM. ('!name = !{ [!n] }' form).

MetaUnnamed !MetaId !MetaExpr
Metadata node declaration. ('!0 = metadata !{ expression }' form).