Llvm.PpLlvm

Description

Pretty print LLVM IR Code.

Top level LLVM objects.

ppLlvmModule :: LlvmModule -> Doc

Print out a whole LLVM module.

ppLlvmComments :: [LMString] -> Doc

Print out a multi-line comment, can be inside a function or on its own

ppLlvmComment :: LMString -> Doc

Print out a comment, can be inside a function or on its own

ppLlvmGlobals :: [LMGlobal] -> Doc

Print out a list of global mutable variable definitions

ppLlvmGlobal :: LMGlobal -> Doc

Print out a global mutable variable definition

ppLlvmAlias :: LlvmAlias -> Doc

Print out an LLVM type alias.

ppLlvmAliases :: [LlvmAlias] -> Doc

Print out a list of LLVM type aliases.

ppLlvmFunctionDecls :: LlvmFunctionDecls -> Doc

Print out a list of function declaration.

ppLlvmFunctionDecl :: LlvmFunctionDecl -> Doc

Print out a function declaration. Declarations define the function type but don't define the actual body of the function.

ppLlvmFunctions :: LlvmFunctions -> Doc

Print out a list of function definitions.

ppLlvmFunction :: LlvmFunction -> Doc

Print out a function definition.

Utility functions

llvmSDoc :: SDoc -> Doc

Convert SDoc to Doc