LlvmCodeGen.Ppr

Description

Pretty print helpers for the LLVM Code generator.

Documentation

pprLlvmHeader :: SDoc

Header code for LLVM modules

pprLlvmCmmDecl :: LlvmCmmDecl -> LlvmM (SDoc, [LlvmVar])

Pretty print LLVM code

pprLlvmData :: LlvmData -> SDoc

Pretty print LLVM data code

infoSection :: String

The section we are putting info tables and their entry code into, should be unique since we process the assembly pattern matching this.