LoadIface

Documentation

loadModuleInterface :: SDoc -> Module -> TcM ModIface

Load interface for a module.

loadModuleInterfaces :: SDoc -> [Module] -> TcM ()

Load interfaces for a collection of modules.

loadSrcInterface :: SDoc -> ModuleName -> IsBootInterface -> Maybe FastString -> RnM ModIface

Load the interface corresponding to an import directive in source code. On a failure, fail in the monad with an error message.

loadInterfaceForName :: SDoc -> Name -> TcRn ModIface

Loads the interface for a given Name.

loadInterface :: SDoc -> Module -> WhereFrom -> IfM lcl (MaybeErr Message ModIface)

loadWiredInHomeIface :: Name -> IfM lcl ()

An IfM function to load the home interface for a wired-in thing, so that we're sure that we see its instance declarations and rules See Note [Loading instances for wired-in things] in TcIface

loadSysInterface :: SDoc -> Module -> IfM lcl ModIface

Loads a system interface and throws an exception if it fails

loadUserInterface :: Bool -> SDoc -> Module -> IfM lcl ModIface

Loads a user interface and throws an exception if it fails. The first parameter indicates whether we should import the boot variant of the module

findAndReadIface :: SDoc -> Module -> IsBootInterface -> TcRnIf gbl lcl (MaybeErr Message (ModIface, FilePath))

readIface :: Module -> FilePath -> IsBootInterface -> TcRnIf gbl lcl (MaybeErr Message ModIface)

loadDecls :: Bool -> [(Fingerprint, IfaceDecl)] -> IfL [(Name, TyThing)]

initExternalPackageState :: ExternalPackageState

ifaceStats :: ExternalPackageState -> SDoc

pprModIface :: ModIface -> SDoc

showIface :: HscEnv -> FilePath -> IO ()

Read binary interface, and print it out