MatchLit

Documentation

dsLit :: HsLit GhcRn -> DsM CoreExpr

dsOverLit :: HsOverLit GhcTc -> DsM CoreExpr

Post-typechecker, the HsExpr field of an OverLit contains (an expression for) the literal value itself.

hsLitKey :: DynFlags -> HsLit GhcTc -> Literal

tidyLitPat :: HsLit GhcTc -> Pat GhcTc

tidyNPat :: HsOverLit GhcTc -> Maybe (SyntaxExpr GhcTc) -> SyntaxExpr GhcTc -> Type -> Pat GhcTc

matchLiterals :: [Id] -> Type -> [[EquationInfo]] -> DsM MatchResult

matchNPlusKPats :: [Id] -> Type -> [EquationInfo] -> DsM MatchResult

matchNPats :: [Id] -> Type -> [EquationInfo] -> DsM MatchResult

warnAboutIdentities :: DynFlags -> CoreExpr -> Type -> DsM ()

warnAboutOverflowedOverLit :: HsOverLit GhcTc -> DsM ()

Emit warnings on overloaded integral literals which overflow the bounds implied by their type.

warnAboutOverflowedLit :: HsLit GhcTc -> DsM ()

Emit warnings on integral literals which overflow the boudns implied by their type.

warnAboutEmptyEnumerations :: DynFlags -> LHsExpr GhcTc -> Maybe (LHsExpr GhcTc) -> LHsExpr GhcTc -> DsM ()

Warns about [2,3 .. 1] which returns the empty list. Only works for integral types, not floating point.