Outputable

Description

This module defines classes and functions for pretty-printing. It also exports a number of helpful debugging and other utilities such as trace and panic.

The interface to this module is very similar to the standard Hughes-PJ pretty printing module, except that it exports a number of additional functions that are rarely used, and works over the SDoc type.

Type classes

class Outputable a where

Class designating that some type has an SDoc representation

Minimal complete definition

Nothing

Methods

ppr :: a -> SDoc

pprPrec :: Rational -> a -> SDoc

class Outputable a => OutputableBndr a where

When we print a binder, we often want to print its type too. The OutputableBndr class encapsulates this idea.

Minimal complete definition

pprPrefixOcc, pprInfixOcc

Methods

pprBndr :: BindingSite -> a -> SDoc

pprPrefixOcc, pprInfixOcc :: a -> SDoc

Pretty printing combinators

data SDoc

runSDoc :: SDoc -> SDocContext -> Doc

initSDocContext :: DynFlags -> PprStyle -> SDocContext

docToSDoc :: Doc -> SDoc

interppSP :: Outputable a => [a] -> SDoc

Returns the separated concatenation of the pretty printed things.

interpp'SP :: Outputable a => [a] -> SDoc

Returns the comma-separated concatenation of the pretty printed things.

pprQuotedList :: Outputable a => [a] -> SDoc

Returns the comma-separated concatenation of the quoted pretty printed things.

[x,y,z]  ==>  `x', `y', `z'

pprWithCommas

Arguments

:: (a -> SDoc)
The pretty printing function to use

-> [a]
The things to be pretty printed

-> SDoc
SDoc where the things have been pretty printed, comma-separated and finally packed into a paragraph.

quotedListWithOr :: [SDoc] -> SDoc

empty :: SDoc

nest :: Int -> SDoc -> SDoc

Indent SDoc some specified amount

char :: Char -> SDoc

text :: String -> SDoc

ftext :: FastString -> SDoc

ptext :: LitString -> SDoc

ztext :: FastZString -> SDoc

int :: Int -> SDoc

intWithCommas :: Integral a => a -> SDoc

integer :: Integer -> SDoc

float :: Float -> SDoc

double :: Double -> SDoc

rational :: Rational -> SDoc

parens :: SDoc -> SDoc

cparen :: Bool -> SDoc -> SDoc

brackets :: SDoc -> SDoc

braces :: SDoc -> SDoc

quotes :: SDoc -> SDoc

quote :: SDoc -> SDoc

doubleQuotes :: SDoc -> SDoc

angleBrackets :: SDoc -> SDoc

paBrackets :: SDoc -> SDoc

semi :: SDoc

comma :: SDoc

colon :: SDoc

dcolon :: SDoc

space :: SDoc

equals :: SDoc

dot :: SDoc

arrow :: SDoc

larrow :: SDoc

darrow :: SDoc

arrowt :: SDoc

larrowt :: SDoc

arrowtt :: SDoc

larrowtt :: SDoc

lparen :: SDoc

rparen :: SDoc

lbrack :: SDoc

rbrack :: SDoc

lbrace :: SDoc

rbrace :: SDoc

underscore :: SDoc

blankLine :: SDoc

forAllLit :: SDoc

(<>) :: SDoc -> SDoc -> SDoc

Join two SDoc together horizontally without a gap

(<+>) :: SDoc -> SDoc -> SDoc

Join two SDoc together horizontally with a gap between them

hcat :: [SDoc] -> SDoc

Concatenate SDoc horizontally

hsep :: [SDoc] -> SDoc

Concatenate SDoc horizontally with a space between each one

($$) :: SDoc -> SDoc -> SDoc

Join two SDoc together vertically; if there is no vertical overlap it "dovetails" the two onto one line

($+$) :: SDoc -> SDoc -> SDoc

Join two SDoc together vertically

vcat :: [SDoc] -> SDoc

Concatenate SDoc vertically with dovetailing

sep :: [SDoc] -> SDoc

Separate: is either like hsep or like vcat, depending on what fits

cat :: [SDoc] -> SDoc

Catenate: is either like hcat or like vcat, depending on what fits

fsep :: [SDoc] -> SDoc

A paragraph-fill combinator. It's much like sep, only it keeps fitting things on one line until it can't fit any more.

fcat :: [SDoc] -> SDoc

This behaves like fsep, but it uses <> for horizontal conposition rather than <+>

hang

Arguments

:: SDoc
The header

-> Int
Amount to indent the hung body

-> SDoc
The hung body, indented and placed below the header

-> SDoc

punctuate

Arguments

:: SDoc
The punctuation

-> [SDoc]
The list that will have punctuation added between every adjacent pair of elements

-> [SDoc]
Punctuated list

ppWhen :: Bool -> SDoc -> SDoc

ppUnless :: Bool -> SDoc -> SDoc

speakNth :: Int -> SDoc

Converts an integer to a verbal index:

speakNth 1 = text "first"
speakNth 5 = text "fifth"
speakNth 21 = text "21st"

speakNTimes :: Int -> SDoc

Converts a strictly positive integer into a number of times:

speakNTimes 1 = text "once"
speakNTimes 2 = text "twice"
speakNTimes 4 = text "4 times"

speakN :: Int -> SDoc

Converts an integer to a verbal multiplicity:

speakN 0 = text "none"
speakN 5 = text "five"
speakN 10 = text "10"

speakNOf :: Int -> SDoc -> SDoc

Converts an integer and object description to a statement about the multiplicity of those objects:

speakNOf 0 (text "melon") = text "no melons"
speakNOf 1 (text "melon") = text "one melon"
speakNOf 3 (text "melon") = text "three melons"

plural :: [a] -> SDoc

Determines the pluralisation suffix appropriate for the length of a list:

plural [] = char 's'
plural ["Hello"] = empty
plural ["Hello", "World"] = char 's'

isOrAre :: [a] -> SDoc

Determines the form of to be appropriate for the length of a list:

isOrAre [] = ptext (sLit "are")
isOrAre ["Hello"] = ptext (sLit "is")
isOrAre ["Hello", "World"] = ptext (sLit "are")

coloured :: PprColour -> SDoc -> SDoc

Apply the given colour/style for the argument.

Only takes effect if colours are enabled.

data PprColour

A colour/style for use with coloured.

colType :: PprColour

colCoerc :: PprColour

colDataCon :: PprColour

colBinder :: PprColour

bold :: SDoc -> SDoc

keyword :: SDoc -> SDoc

Converting SDoc into strings and outputing it

printForC :: DynFlags -> Handle -> SDoc -> IO ()

printForAsm :: DynFlags -> Handle -> SDoc -> IO ()

printForUser :: DynFlags -> Handle -> PrintUnqualified -> SDoc -> IO ()

printForUserPartWay :: DynFlags -> Handle -> Int -> PrintUnqualified -> SDoc -> IO ()

pprCode :: CodeStyle -> SDoc -> SDoc

mkCodeStyle :: CodeStyle -> PprStyle

showSDoc :: DynFlags -> SDoc -> String

showSDocUnsafe :: SDoc -> String

showSDocOneLine :: DynFlags -> SDoc -> String

showSDocForUser :: DynFlags -> PrintUnqualified -> SDoc -> String

showSDocDebug :: DynFlags -> SDoc -> String

showSDocDump :: DynFlags -> SDoc -> String

showSDocDumpOneLine :: DynFlags -> SDoc -> String

showSDocUnqual :: DynFlags -> SDoc -> String

showPpr :: Outputable a => DynFlags -> a -> String

renderWithStyle :: DynFlags -> SDoc -> PprStyle -> String

pprInfixVar :: Bool -> SDoc -> SDoc

pprPrefixVar :: Bool -> SDoc -> SDoc

pprHsChar :: Char -> SDoc

Special combinator for showing character literals.

pprHsString :: FastString -> SDoc

Special combinator for showing string literals.

pprHsBytes :: ByteString -> SDoc

Special combinator for showing string literals.

pprFastFilePath :: FastString -> SDoc

Controlling the style in which output is printed

data BindingSite

BindingSite is used to tell the thing that prints binder what language construct is binding the identifier. This can be used to decide how much info to print.

Constructors

LambdaBind

CaseBind

LetBind

data PprStyle

data CodeStyle

Constructors

CStyle

AsmStyle

data PrintUnqualified

When printing code that contains original names, we need to map the original names back to something the user understands. This is the purpose of the triple of functions that gets passed around when rendering SDoc.

Constructors

QueryQualify

Fields

queryQualifyName :: QueryQualifyName

queryQualifyModule :: QueryQualifyModule

queryQualifyPackage :: QueryQualifyPackage

type QueryQualifyName = Module -> OccName -> QualifyName

given an original name, this function tells you which module name it should be qualified with when printing for the user, if any. For example, given Control.Exception.catch, which is in scope as Exception.catch, this fuction will return Just Exception. Note that the return value is a ModuleName, not a Module, because in source code, names are qualified by ModuleNames.

type QueryQualifyModule = Module -> Bool

For a given module, we need to know whether to print it with a package name to disambiguate it.

type QueryQualifyPackage = PackageKey -> Bool

For a given package, we need to know whether to print it with the package key to disambiguate it.

reallyAlwaysQualify :: PrintUnqualified

reallyAlwaysQualifyNames :: QueryQualifyName

alwaysQualify :: PrintUnqualified

alwaysQualifyNames :: QueryQualifyName

NB: This won't ever show package IDs

alwaysQualifyModules :: QueryQualifyModule

neverQualify :: PrintUnqualified

neverQualifyNames :: QueryQualifyName

neverQualifyModules :: QueryQualifyModule

data QualifyName

Constructors

NameUnqual

NameQual ModuleName

NameNotInScope1

NameNotInScope2

queryQual :: PprStyle -> PrintUnqualified

sdocWithDynFlags :: (DynFlags -> SDoc) -> SDoc

sdocWithPlatform :: (Platform -> SDoc) -> SDoc

getPprStyle :: (PprStyle -> SDoc) -> SDoc

withPprStyle :: PprStyle -> SDoc -> SDoc

withPprStyleDoc :: DynFlags -> PprStyle -> SDoc -> Doc

pprDeeper :: SDoc -> SDoc

pprDeeperList :: ([SDoc] -> SDoc) -> [SDoc] -> SDoc

pprSetDepth :: Depth -> SDoc -> SDoc

codeStyle :: PprStyle -> Bool

userStyle :: PprStyle -> Bool

debugStyle :: PprStyle -> Bool

dumpStyle :: PprStyle -> Bool

asmStyle :: PprStyle -> Bool

ifPprDebug :: SDoc -> SDoc

qualName :: PprStyle -> QueryQualifyName

qualModule :: PprStyle -> QueryQualifyModule

qualPackage :: PprStyle -> QueryQualifyPackage

mkErrStyle :: DynFlags -> PrintUnqualified -> PprStyle

Style for printing error messages

defaultErrStyle :: DynFlags -> PprStyle

defaultDumpStyle :: PprStyle

mkDumpStyle :: PrintUnqualified -> PprStyle

defaultUserStyle :: PprStyle

mkUserStyle :: PrintUnqualified -> Depth -> PprStyle

cmdlineParserStyle :: PprStyle

data Depth

Constructors

AllTheWay

PartWay Int

Error handling and debugging utilities

pprPanic :: String -> SDoc -> a

Throw an exception saying "bug in GHC"

pprSorry :: String -> SDoc -> a

Throw an exception saying "this isn't finished yet"

assertPprPanic :: String -> Int -> SDoc -> a

Panic with an assertation failure, recording the given file and line number. Should typically be accessed with the ASSERT family of macros

pprPanicFastInt :: String -> SDoc -> FastInt

Specialization of pprPanic that can be safely used with FastInt

pprPgmError :: String -> SDoc -> a

Throw an exception saying "bug in pgm being compiled" (used for unusual program errors)

pprTrace :: String -> SDoc -> a -> a

If debug output is on, show some SDoc on the screen

warnPprTrace :: Bool -> String -> Int -> SDoc -> a -> a

Just warn about an assertion failure, recording the given file and line number. Should typically be accessed with the WARN macros

trace :: String -> a -> a

The trace function outputs the trace message given as its first argument, before returning the second argument as its result.

For example, this returns the value of f x but first outputs the message.

trace ("calling f with x = " ++ show x) (f x)

The trace function should only be used for debugging, or for monitoring execution. The function is not referentially transparent: its type indicates that it is a pure function but it has the side effect of outputting the trace message.

pgmError :: String -> a

Panics and asserts.

panic :: String -> a

Panics and asserts.

sorry :: String -> a

Panics and asserts.

panicFastInt :: String -> FastInt

Panic while pretending to return an unboxed int. You can't use the regular panic functions in expressions producing unboxed ints because they have the wrong kind.

assertPanic :: String -> Int -> a

Throw an failed assertion exception for a given filename and line number.

pprDebugAndThen :: DynFlags -> (String -> a) -> SDoc -> SDoc -> a