PPC.Regs

Documentation

virtualRegSqueeze :: RegClass -> VirtualReg -> FastInt

regSqueeze_class reg Calculuate the maximum number of register colors that could be denied to a node of this class due to having this reg as a neighbour.

realRegSqueeze :: RegClass -> RealReg -> FastInt

mkVirtualReg :: Unique -> Size -> VirtualReg

regDotColor :: RealReg -> SDoc

data Imm

Constructors

ImmInt Int

ImmInteger Integer

ImmCLbl CLabel

ImmLit Doc

ImmIndex CLabel Int

ImmFloat Rational

ImmDouble Rational

ImmConstantSum Imm Imm

ImmConstantDiff Imm Imm

LO Imm

HI Imm

HA Imm

strImmLit :: String -> Imm

litToImm :: CmmLit -> Imm

data AddrMode

Constructors

AddrRegReg Reg Reg

AddrRegImm Reg Imm

addrOffset :: AddrMode -> Int -> Maybe AddrMode

spRel :: Int -> AddrMode

argRegs :: RegNo -> [Reg]

allArgRegs :: [Reg]

callClobberedRegs :: [Reg]

allMachRegNos :: [RegNo]

classOfRealReg :: RealReg -> RegClass

showReg :: RegNo -> String

allFPArgRegs :: [Reg]

fits16Bits :: Integral a => a -> Bool

makeImmediate :: Integral a => Width -> Bool -> a -> Maybe Imm

fReg :: Int -> RegNo

sp :: Reg

r3 :: Reg

r4 :: Reg

r27 :: Reg

r28 :: Reg

f1 :: Reg

f20 :: Reg

f21 :: Reg

freeReg :: RegNo -> FastBool

globalRegMaybe :: GlobalReg -> Maybe Reg

allocatableRegs :: [RealReg]