PackageConfig

Description

Package configuration information: essentially the interface to Cabal, with some utilities

(c) The University of Glasgow, 2004

Documentation

Mostly the compiler deals in terms of PackageKeys, which are md5 hashes of a package ID, keys of its dependencies, and Cabal flags. You're expected to pass in the package key in the -this-package-key flag. However, for wired-in packages like base & rts, we don't necessarily know what the version is, so these are handled specially; see .

PackageKey

packageConfigId :: PackageConfig -> PackageKey

Get the GHC PackageKey right out of a Cabalish PackageConfig

The PackageConfig type: information about a package

type PackageConfig = InstalledPackageInfo InstalledPackageId SourcePackageId PackageName PackageKey ModuleName

data InstalledPackageInfo instpkgid srcpkgid srcpkgname pkgkey modulename :: * -> * -> * -> * -> * -> *

This is a subset of Cabal's InstalledPackageInfo, with just the bits that GHC is interested in.

Constructors

InstalledPackageInfo

Fields

installedPackageId :: instpkgid

sourcePackageId :: srcpkgid

packageName :: srcpkgname

packageVersion :: Version

packageKey :: pkgkey

depends :: [instpkgid]

importDirs :: [FilePath]

hsLibraries :: [String]

extraLibraries :: [String]

extraGHCiLibraries :: [String]

libraryDirs :: [FilePath]

frameworks :: [String]

frameworkDirs :: [FilePath]

ldOptions :: [String]

ccOptions :: [String]

includes :: [String]

includeDirs :: [FilePath]

haddockInterfaces :: [FilePath]

haddockHTMLs :: [FilePath]

exposedModules :: [ExposedModule instpkgid modulename]

hiddenModules :: [modulename]

instantiatedWith :: [(modulename, OriginalModule instpkgid modulename)]

exposed :: Bool

trusted :: Bool

newtype InstalledPackageId

Constructors

InstalledPackageId FastString

newtype SourcePackageId

Constructors

SourcePackageId FastString

newtype PackageName

Constructors

PackageName FastString

data Version :: *

A Version represents the version of a software entity.

An instance of Eq is provided, which implements exact equality modulo reordering of the tags in the versionTags field.

An instance of Ord is also provided, which gives lexicographic ordering on the versionBranch fields (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.). This is expected to be sufficient for many uses, but note that you may need to use a more specific ordering for your versioning scheme. For example, some versioning schemes may include pre-releases which have tags "pre1", "pre2", and so on, and these would need to be taken into account when determining ordering. In some cases, date ordering may be more appropriate, so the application would have to look for date tags in the versionTags field and compare those. The bottom line is, don't always assume that compare and other Ord operations are the right thing for every Version.

Similarly, concrete representations of versions may differ. One possible concrete representation is provided (see showVersion and parseVersion), but depending on the application a different concrete representation may be more appropriate.

Constructors

Version

Fields

versionBranch :: [Int]

The numeric branch for this version. This reflects the fact that most software versions are tree-structured; there is a main trunk which is tagged with versions at various points (1,2,3...), and the first branch off the trunk after version 3 is 3.1, the second branch off the trunk after version 3 is 3.2, and so on. The tree can be branched arbitrarily, just by adding more digits.

We represent the branch as a list of Int, so version 3.2.1 becomes [3,2,1]. Lexicographic ordering (i.e. the default instance of Ord for [Int]) gives the natural ordering of branches.

versionTags :: [String]

A version can be tagged with an arbitrary list of strings. The interpretation of the list of tags is entirely dependent on the entity that this version applies to.

defaultPackageConfig :: PackageConfig

installedPackageIdString :: PackageConfig -> String

sourcePackageIdString :: PackageConfig -> String

packageNameString :: PackageConfig -> String

pprPackageConfig :: PackageConfig -> SDoc