PrelNames

Documentation

data Unique

The type of unique identifiers that are used in many places in GHC for fast ordering and equality tests. You should generate these with the functions from the UniqSupply module

class Uniquable a where

Class of things that we can obtain a Unique from

Methods

getUnique :: a -> Unique

hasKey :: Uniquable a => a -> Unique -> Bool

itName :: Unique -> SrcSpan -> Name

mkUnboundName :: RdrName -> Name

isUnboundName :: Name -> Bool

basicKnownKeyNames :: [Name]

genericTyConNames :: [Name]

pRELUDE :: Module

gHC_PRIM :: Module

cONTROL_EXCEPTION_BASE :: Module

gHC_EXTS :: Module

rANDOM :: Module

gHC_DESUGAR :: Module

cONTROL_APPLICATIVE :: Module

aRROW :: Module

mONAD_ZIP :: Module

mONAD_FIX :: Module

mONAD :: Module

gHC_WORD :: Module

gHC_INT :: Module

lEX :: Module

rEAD_PREC :: Module

dOTNET :: Module

gENERICS :: Module

tYPEABLE_INTERNAL :: Module

tYPEABLE :: Module

dYNAMIC :: Module

sYSTEM_IO :: Module

gHC_TOP_HANDLER :: Module

gHC_FLOAT :: Module

gHC_REAL :: Module

gHC_ERR :: Module

gHC_PTR :: Module

gHC_STABLE :: Module

gHC_ARR :: Module

gHC_ST :: Module

gHC_IO_Exception :: Module

gHC_IO :: Module

gHC_CONC :: Module

dATA_TRAVERSABLE :: Module

dATA_FOLDABLE :: Module

dATA_STRING :: Module

dATA_EITHER :: Module

dATA_TUPLE :: Module

gHC_TUPLE :: Module

gHC_LIST :: Module

gHC_INTEGER_TYPE :: Module

gHC_NUM :: Module

gHC_READ :: Module

gHC_SHOW :: Module

gHC_CSTRING :: Module

gHC_ENUM :: Module

gHC_BASE :: Module

gHC_CLASSES :: Module

gHC_MAGIC :: Module

gHC_GENERICS :: Module

gHC_TYPES :: Module

gHC_PARR' :: Module

mAIN :: Module

rOOT_MAIN :: Module

iNTERACTIVE :: Module

pRELUDE_NAME :: ModuleName

mAIN_NAME :: ModuleName

dATA_ARRAY_PARALLEL_NAME :: ModuleName

dATA_ARRAY_PARALLEL_PRIM_NAME :: ModuleName

mkPrimModule :: FastString -> Module

mkIntegerModule :: FastString -> Module

mkBaseModule :: FastString -> Module

mkBaseModule_ :: ModuleName -> Module

mkThisGhcModule :: FastString -> Module

mkThisGhcModule_ :: ModuleName -> Module

mkMainModule :: FastString -> Module

mkMainModule_ :: ModuleName -> Module

mkTupleModule :: TupleSort -> Arity -> Module

main_RDR_Unqual :: RdrName

forall_tv_RDR :: RdrName

dot_tv_RDR :: RdrName

eq_RDR :: RdrName

gtTag_RDR :: RdrName

eqTag_RDR :: RdrName

ltTag_RDR :: RdrName

compare_RDR :: RdrName

gt_RDR :: RdrName

lt_RDR :: RdrName

le_RDR :: RdrName

ne_RDR :: RdrName

ge_RDR :: RdrName

eqClass_RDR :: RdrName

monadClass_RDR :: RdrName

enumClass_RDR :: RdrName

ordClass_RDR :: RdrName

numClass_RDR :: RdrName

map_RDR :: RdrName

append_RDR :: RdrName

foldr_RDR :: RdrName

failM_RDR :: RdrName

bindM_RDR :: RdrName

returnM_RDR :: RdrName

build_RDR :: RdrName

left_RDR :: RdrName

right_RDR :: RdrName

fromEnum_RDR :: RdrName

toEnum_RDR :: RdrName

enumFrom_RDR :: RdrName

enumFromThenTo_RDR :: RdrName

enumFromThen_RDR :: RdrName

enumFromTo_RDR :: RdrName

ratioDataCon_RDR :: RdrName

timesInteger_RDR :: RdrName

plusInteger_RDR :: RdrName

ioDataCon_RDR :: RdrName

eqString_RDR :: RdrName

unpackCStringUtf8_RDR :: RdrName

unpackCStringFoldr_RDR :: RdrName

unpackCString_RDR :: RdrName

newStablePtr_RDR :: RdrName

wordDataCon_RDR :: RdrName

bindIO_RDR :: RdrName

returnIO_RDR :: RdrName

fromInteger_RDR :: RdrName

plus_RDR :: RdrName

times_RDR :: RdrName

minus_RDR :: RdrName

fromRational_RDR :: RdrName

fromString_RDR :: RdrName

compose_RDR :: RdrName

not_RDR :: RdrName

unsafeRangeSize_RDR :: RdrName

unsafeIndex_RDR :: RdrName

index_RDR :: RdrName

inRange_RDR :: RdrName

range_RDR :: RdrName

and_RDR :: RdrName

maxBound_RDR :: RdrName

minBound_RDR :: RdrName

pred_RDR :: RdrName

succ_RDR :: RdrName

getTag_RDR :: RdrName

readList_RDR :: RdrName

lexP_RDR :: RdrName

choose_RDR :: RdrName

parens_RDR :: RdrName

readPrec_RDR :: RdrName

readListPrecDefault_RDR :: RdrName

readListPrec_RDR :: RdrName

readListDefault_RDR :: RdrName

punc_RDR :: RdrName

symbol_RDR :: RdrName

ident_RDR :: RdrName

step_RDR :: RdrName

prec_RDR :: RdrName

reset_RDR :: RdrName

alt_RDR :: RdrName

showList_RDR :: RdrName

showParen_RDR :: RdrName

showSpace_RDR :: RdrName

showString_RDR :: RdrName

showsPrec_RDR :: RdrName

showList___RDR :: RdrName

typeOf_RDR :: RdrName

mkTyConApp_RDR :: RdrName

mkTyCon_RDR :: RdrName

undefined_RDR :: RdrName

error_RDR :: RdrName

u1DataCon_RDR :: RdrName

notAssocDataCon_RDR :: RdrName

rightAssocDataCon_RDR :: RdrName

leftAssocDataCon_RDR :: RdrName

infixDataCon_RDR :: RdrName

prefixDataCon_RDR :: RdrName

selName_RDR :: RdrName

arityDataCon_RDR :: RdrName

noArityDataCon_RDR :: RdrName

conIsRecord_RDR :: RdrName

conFixity_RDR :: RdrName

conName_RDR :: RdrName

moduleName_RDR :: RdrName

datatypeName_RDR :: RdrName

to1_RDR :: RdrName

to_RDR :: RdrName

from1_RDR :: RdrName

from_RDR :: RdrName

comp1DataCon_RDR :: RdrName

prodDataCon_RDR :: RdrName

r1DataCon_RDR :: RdrName

l1DataCon_RDR :: RdrName

m1DataCon_RDR :: RdrName

k1DataCon_RDR :: RdrName

rec1DataCon_RDR :: RdrName

par1DataCon_RDR :: RdrName

fmap_RDR :: RdrName

traverse_RDR :: RdrName

foldable_foldr_RDR :: RdrName

ap_RDR :: RdrName

pure_RDR :: RdrName

varQual_RDR :: Module -> FastString -> RdrName

dataQual_RDR :: Module -> FastString -> RdrName

clsQual_RDR :: Module -> FastString -> RdrName

tcQual_RDR :: Module -> FastString -> RdrName

wildCardName :: Name

runMainIOName :: Name

orderingTyConName :: Name

gtDataConName :: Name

eqDataConName :: Name

ltDataConName :: Name

eitherTyConName :: Name

rightDataConName :: Name

leftDataConName :: Name

v1TyConName :: Name

rep1TyConName :: Name

repTyConName :: Name

noSelTyConName :: Name

s1TyConName :: Name

c1TyConName :: Name

d1TyConName :: Name

par0TyConName :: Name

rec0TyConName :: Name

sTyConName :: Name

cTyConName :: Name

dTyConName :: Name

pTyConName :: Name

rTyConName :: Name

compTyConName :: Name

prodTyConName :: Name

sumTyConName :: Name

m1TyConName :: Name

k1TyConName :: Name

rec1TyConName :: Name

par1TyConName :: Name

u1TyConName :: Name

unpackCStringName :: Name

stringTyConName :: Name

eqStringName :: Name

unpackCStringUtf8Name :: Name

unpackCStringFoldrName :: Name

inlineIdName :: Name

undefinedName :: Name

fmapName :: Name

functorClassName :: Name

geName :: Name

ordClassName :: Name

eqName :: Name

eqClassName :: Name

monadClassName :: Name

failMName :: Name

returnMName :: Name

bindMName :: Name

thenMName :: Name

applicativeClassName :: Name

traversableClassName :: Name

foldableClassName :: Name

groupWithName :: Name

fromStringName :: Name

opaqueTyConName :: Name

dollarName :: Name

breakpointAutoName :: Name

breakpointCondName :: Name

breakpointName :: Name

assertName :: Name

appendName :: Name

mapName :: Name

augmentName :: Name

buildName :: Name

foldrName :: Name

otherwiseIdName :: Name

breakpointJumpName :: Name

breakpointCondJumpName :: Name

breakpointAutoJumpName :: Name

fstName :: Name

sndName :: Name

numClassName :: Name

negateName :: Name

minusName :: Name

fromIntegerName :: Name

integerTyConName :: Name

shiftRIntegerName :: Name

shiftLIntegerName :: Name

complementIntegerName :: Name

xorIntegerName :: Name

orIntegerName :: Name

andIntegerName :: Name

lcmIntegerName :: Name

gcdIntegerName :: Name

encodeDoubleIntegerName :: Name

encodeFloatIntegerName :: Name

doubleFromIntegerName :: Name

floatFromIntegerName :: Name

remIntegerName :: Name

quotIntegerName :: Name

divModIntegerName :: Name

quotRemIntegerName :: Name

compareIntegerName :: Name

geIntegerName :: Name

ltIntegerName :: Name

gtIntegerName :: Name

leIntegerName :: Name

signumIntegerName :: Name

absIntegerName :: Name

neqIntegerName :: Name

eqIntegerName :: Name

negateIntegerName :: Name

minusIntegerName :: Name

integerToIntName :: Name

integerToWordName :: Name

smallIntegerName :: Name

timesIntegerName :: Name

plusIntegerName :: Name

integerToInt64Name :: Name

integerToWord64Name :: Name

mkIntegerName :: Name

rationalTyConName :: Name

realToFracName :: Name

fromIntegralName :: Name

toRationalName :: Name

toIntegerName :: Name

fromRationalName :: Name

fractionalClassName :: Name

realFracClassName :: Name

integralClassName :: Name

realClassName :: Name

ratioDataConName :: Name

ratioTyConName :: Name

floatingClassName :: Name

realFloatClassName :: Name

ixClassName :: Name

typeableClassName :: Name

typeable7ClassName :: Name

typeable6ClassName :: Name

typeable5ClassName :: Name

typeable4ClassName :: Name

typeable3ClassName :: Name

typeable2ClassName :: Name

typeable1ClassName :: Name

typeableClassNames :: [Name]

dataClassName :: Name

assertErrorName :: Name

enumClassName :: Name

boundedClassName :: Name

enumFromThenToName :: Name

enumFromThenName :: Name

enumFromToName :: Name

enumFromName :: Name

concatName :: Name

zipName :: Name

filterName :: Name

showClassName :: Name

readClassName :: Name

genClassName :: Name

selectorClassName :: Name

constructorClassName :: Name

datatypeClassName :: Name

gen1ClassName :: Name

ioTyConName :: Name

failIOName :: Name

returnIOName :: Name

bindIOName :: Name

thenIOName :: Name

ioDataConName :: Name

printName :: Name

int8TyConName :: Name

int64TyConName :: Name

int32TyConName :: Name

int16TyConName :: Name

word8TyConName :: Name

wordDataConName :: Name

wordTyConName :: Name

word64TyConName :: Name

word32TyConName :: Name

word16TyConName :: Name

ptrTyConName :: Name

funPtrTyConName :: Name

stablePtrTyConName :: Name

newStablePtrName :: Name

runSTRepName :: Name

monadFixClassName :: Name

mfixName :: Name

arrAName :: Name

loopAName :: Name

choiceAName :: Name

appAName :: Name

firstAName :: Name

composeAName :: Name

guardMName :: Name

mzipName :: Name

liftMName :: Name

toAnnotationWrapperName :: Name

monadPlusClassName :: Name

isStringClassName :: Name

randomGenClassName :: Name

randomClassName :: Name

objectTyConName :: Name

unmarshalObjectName :: Name

checkDotnetResName :: Name

unmarshalStringName :: Name

marshalStringName :: Name

marshalObjectName :: Name

cORE_MONAD :: Module

pluginTyConName :: Name

varQual :: Module -> FastString -> Unique -> Name

clsQual :: Module -> FastString -> Unique -> Name

tcQual :: Module -> FastString -> Unique -> Name

mk_known_key_name :: NameSpace -> Module -> FastString -> Unique -> Name

conName :: Module -> FastString -> Unique -> Name

methName :: Module -> FastString -> Unique -> Name

boundedClassKey :: Unique

ixClassKey :: Unique

showClassKey :: Unique

realFracClassKey :: Unique

realFloatClassKey :: Unique

realClassKey :: Unique

readClassKey :: Unique

ordClassKey :: Unique

numClassKey :: Unique

functorClassKey :: Unique

dataClassKey :: Unique

monadClassKey :: Unique

integralClassKey :: Unique

fractionalClassKey :: Unique

floatingClassKey :: Unique

eqClassKey :: Unique

enumClassKey :: Unique

typeableClassKey :: Unique

typeable7ClassKey :: Unique

typeable6ClassKey :: Unique

typeable5ClassKey :: Unique

typeable4ClassKey :: Unique

typeable3ClassKey :: Unique

typeable2ClassKey :: Unique

typeable1ClassKey :: Unique

monadFixClassKey :: Unique

monadPlusClassKey :: Unique

randomGenClassKey :: Unique

randomClassKey :: Unique

isStringClassKey :: Unique

applicativeClassKey :: Unique

traversableClassKey :: Unique

foldableClassKey :: Unique

genClassKey :: Unique

selectorClassKey :: Unique

constructorClassKey :: Unique

datatypeClassKey :: Unique

gen1ClassKey :: Unique

addrPrimTyConKey :: Unique

eqTyConKey :: Unique

anyTyConKey :: Unique

stablePtrTyConKey :: Unique

stablePtrPrimTyConKey :: Unique

realWorldTyConKey :: Unique

rationalTyConKey :: Unique

ratioTyConKey :: Unique

mVarPrimTyConKey :: Unique

orderingTyConKey :: Unique

mutableByteArrayPrimTyConKey :: Unique

mutableArrayArrayPrimTyConKey :: Unique

mutableArrayPrimTyConKey :: Unique

weakPrimTyConKey :: Unique

foreignObjPrimTyConKey :: Unique

listTyConKey :: Unique

digitsTyConKey :: Unique

integerTyConKey :: Unique

int64TyConKey :: Unique

int64PrimTyConKey :: Unique

int32TyConKey :: Unique

int32PrimTyConKey :: Unique

int16TyConKey :: Unique

int8TyConKey :: Unique

intTyConKey :: Unique

intPrimTyConKey :: Unique

funTyConKey :: Unique

floatTyConKey :: Unique

floatPrimTyConKey :: Unique

doubleTyConKey :: Unique

doublePrimTyConKey :: Unique

charTyConKey :: Unique

charPrimTyConKey :: Unique

byteArrayPrimTyConKey :: Unique

boolTyConKey :: Unique

arrayArrayPrimTyConKey :: Unique

arrayPrimTyConKey :: Unique

statePrimTyConKey :: Unique

eqPrimTyConKey :: Unique

tVarPrimTyConKey :: Unique

funPtrTyConKey :: Unique

ptrTyConKey :: Unique

bcoPrimTyConKey :: Unique

threadIdPrimTyConKey :: Unique

typeConKey :: Unique

boxityConKey :: Unique

kindConKey :: Unique

anyBoxConKey :: Unique

unliftedConKey :: Unique

liftedConKey :: Unique

word64TyConKey :: Unique

word64PrimTyConKey :: Unique

word32TyConKey :: Unique

word32PrimTyConKey :: Unique

word16TyConKey :: Unique

word8TyConKey :: Unique

wordTyConKey :: Unique

wordPrimTyConKey :: Unique

ioTyConKey :: Unique

mutVarPrimTyConKey :: Unique

stableNameTyConKey :: Unique

stableNamePrimTyConKey :: Unique

parrTyConKey :: Unique

objectTyConKey :: Unique

eitherTyConKey :: Unique

tySuperKindTyConKey :: Unique

liftedTypeKindTyConKey :: Unique

constraintKindTyConKey :: Unique

argTypeKindTyConKey :: Unique

ubxTupleKindTyConKey :: Unique

unliftedTypeKindTyConKey :: Unique

openTypeKindTyConKey :: Unique

anyKindTyConKey :: Unique

symCoercionTyConKey :: Unique

cselRCoercionTyConKey :: Unique

csel2CoercionTyConKey :: Unique

csel1CoercionTyConKey :: Unique

unsafeCoercionTyConKey :: Unique

instCoercionTyConKey :: Unique

rightCoercionTyConKey :: Unique

leftCoercionTyConKey :: Unique

transCoercionTyConKey :: Unique

pluginTyConKey :: Unique

unknownTyConKey :: Unique

opaqueTyConKey :: Unique

unknown3TyConKey :: Unique

unknown2TyConKey :: Unique

unknown1TyConKey :: Unique

stringTyConKey :: Unique

v1TyConKey :: Unique

rep1TyConKey :: Unique

repTyConKey :: Unique

noSelTyConKey :: Unique

s1TyConKey :: Unique

c1TyConKey :: Unique

d1TyConKey :: Unique

par0TyConKey :: Unique

rec0TyConKey :: Unique

sTyConKey :: Unique

cTyConKey :: Unique

dTyConKey :: Unique

pTyConKey :: Unique

rTyConKey :: Unique

compTyConKey :: Unique

prodTyConKey :: Unique

sumTyConKey :: Unique

m1TyConKey :: Unique

k1TyConKey :: Unique

rec1TyConKey :: Unique

par1TyConKey :: Unique

u1TyConKey :: Unique

unitTyConKey :: Unique

charDataConKey :: Unique

eqBoxDataConKey :: Unique

integerDataConKey :: Unique

ioDataConKey :: Unique

wordDataConKey :: Unique

trueDataConKey :: Unique

stableNameDataConKey :: Unique

ratioDataConKey :: Unique

nilDataConKey :: Unique

intDataConKey :: Unique

floatDataConKey :: Unique

falseDataConKey :: Unique

doubleDataConKey :: Unique

consDataConKey :: Unique

crossDataConKey :: Unique

genUnitDataConKey :: Unique

inrDataConKey :: Unique

inlDataConKey :: Unique

parrDataConKey :: Unique

leftDataConKey :: Unique

rightDataConKey :: Unique

ltDataConKey :: Unique

gtDataConKey :: Unique

eqDataConKey :: Unique

integerGmpSDataConKey :: Unique

integerGmpJDataConKey :: Unique

wildCardKey :: Unique

unpackCStringIdKey :: Unique

unpackCStringFoldrIdKey :: Unique

unpackCStringAppendIdKey :: Unique

unpackCStringUtf8IdKey :: Unique

recConErrorIdKey :: Unique

realWorldPrimIdKey :: Unique

patErrorIdKey :: Unique

runtimeErrorIdKey :: Unique

nonExhaustiveGuardsErrorIdKey :: Unique

noMethodBindingErrorIdKey :: Unique

eqStringIdKey :: Unique

irrefutPatErrorIdKey :: Unique

seqIdKey :: Unique

recSelErrorIdKey :: Unique

foldrIdKey :: Unique

errorIdKey :: Unique

buildIdKey :: Unique

appendIdKey :: Unique

augmentIdKey :: Unique

absentErrorIdKey :: Unique

unsafeCoerceIdKey :: Unique

runSTRepIdKey :: Unique

assertIdKey :: Unique

otherwiseIdKey :: Unique

sndIdKey :: Unique

fstIdKey :: Unique

voidArgIdKey :: Unique

nullAddrIdKey :: Unique

failIOIdKey :: Unique

printIdKey :: Unique

newStablePtrIdKey :: Unique

returnIOIdKey :: Unique

bindIOIdKey :: Unique

zipIdKey :: Unique

filterIdKey :: Unique

concatIdKey :: Unique

mkIntegerIdKey :: Unique

shiftRIntegerIdKey :: Unique

shiftLIntegerIdKey :: Unique

complementIntegerIdKey :: Unique

xorIntegerIdKey :: Unique

orIntegerIdKey :: Unique

andIntegerIdKey :: Unique

lcmIntegerIdKey :: Unique

gcdIntegerIdKey :: Unique

encodeDoubleIntegerIdKey :: Unique

encodeFloatIntegerIdKey :: Unique

doubleFromIntegerIdKey :: Unique

floatFromIntegerIdKey :: Unique

remIntegerIdKey :: Unique

quotIntegerIdKey :: Unique

divModIntegerIdKey :: Unique

quotRemIntegerIdKey :: Unique

compareIntegerIdKey :: Unique

geIntegerIdKey :: Unique

ltIntegerIdKey :: Unique

gtIntegerIdKey :: Unique

leIntegerIdKey :: Unique

signumIntegerIdKey :: Unique

absIntegerIdKey :: Unique

neqIntegerIdKey :: Unique

eqIntegerIdKey :: Unique

negateIntegerIdKey :: Unique

minusIntegerIdKey :: Unique

timesIntegerIdKey :: Unique

plusIntegerIdKey :: Unique

integerToInt64IdKey :: Unique

integerToWord64IdKey :: Unique

integerToIntIdKey :: Unique

integerToWordIdKey :: Unique

smallIntegerIdKey :: Unique

rootMainKey :: Unique

runMainKey :: Unique

thenIOIdKey :: Unique

assertErrorIdKey :: Unique

lazyIdKey :: Unique

breakpointIdKey :: Unique

breakpointAutoJumpIdKey :: Unique

breakpointCondJumpIdKey :: Unique

breakpointJumpIdKey :: Unique

breakpointAutoIdKey :: Unique

breakpointCondIdKey :: Unique

inlineIdKey :: Unique

mapIdKey :: Unique

dollarIdKey :: Unique

groupWithIdKey :: Unique

coercionTokenIdKey :: Unique

unmarshalObjectIdKey :: Unique

checkDotnetResNameIdKey :: Unique

unmarshalStringIdKey :: Unique

marshalStringIdKey :: Unique

marshalObjectIdKey :: Unique

undefinedKey :: Unique

unboundKey :: Unique

fromIntegerClassOpKey :: Unique

fmapClassOpKey :: Unique

returnMClassOpKey :: Unique

thenMClassOpKey :: Unique

bindMClassOpKey :: Unique

failMClassOpKey :: Unique

negateClassOpKey :: Unique

geClassOpKey :: Unique

eqClassOpKey :: Unique

enumFromThenToClassOpKey :: Unique

enumFromToClassOpKey :: Unique

enumFromThenClassOpKey :: Unique

enumFromClassOpKey :: Unique

fromRationalClassOpKey :: Unique

minusClassOpKey :: Unique

mfixIdKey :: Unique

arrAIdKey :: Unique

loopAIdKey :: Unique

choiceAIdKey :: Unique

appAIdKey :: Unique

firstAIdKey :: Unique

composeAIdKey :: Unique

fromStringClassOpKey :: Unique

toAnnotationWrapperIdKey :: Unique

fromIntegralIdKey :: Unique

toRationalClassOpKey :: Unique

toIntegerClassOpKey :: Unique

realToFracIdKey :: Unique

guardMIdKey :: Unique

mzipIdKey :: Unique

liftMIdKey :: Unique

kindKeys :: [Unique]

numericClassKeys :: [Unique]

fractionalClassKeys :: [Unique]

standardClassKeys :: [Unique]

derivableClassKeys :: [Unique]