RegAlloc.Graph.ArchX86

Description

A description of the register set of the X86.

This isn't used directly in GHC proper.

See RegArchBase.hs for the reference. See MachRegs.hs for the actual trivColorable function used in GHC.

Documentation

classOfReg :: Reg -> RegClass

Determine the class of a register

regsOfClass :: RegClass -> UniqSet Reg

Determine all the regs that make up a certain class.

regName :: Reg -> Maybe String

Determine the common name of a reg returns Nothing if this reg is not part of the machine.

regAlias :: Reg -> UniqSet Reg

Which regs alias what other regs.

worst :: Int -> RegClass -> RegClass -> Int

Optimised versions of RegColorBase.{worst, squeese} specific to x86

squeese :: RegClass -> [(Int, RegClass)] -> Int