RegAlloc.Graph.Main

Description

Graph coloring register allocator.

Documentation

regAlloc

Arguments

:: (Outputable statics, Outputable instr, Instruction instr)

=> DynFlags

-> UniqFM (UniqSet RealReg)
registers we can use for allocation

-> UniqSet Int
set of available spill slots.

-> [LiveCmmDecl statics instr]
code annotated with liveness information.

-> UniqSM ([NatCmmDecl statics instr], [RegAllocStats statics instr])
code with registers allocated and stats for each stage of allocation

The top level of the graph coloring register allocator.