RnExpr

Documentation

rnLExpr :: LHsExpr RdrName -> RnM (LHsExpr Name, FreeVars)

rnExpr :: HsExpr RdrName -> RnM (HsExpr Name, FreeVars)

rnStmts

Arguments

:: Outputable (body RdrName)

=> HsStmtContext Name

-> (Located (body RdrName) -> RnM (Located (body Name), FreeVars))
How to rename the body of each statement (e.g. rnLExpr)

-> [LStmt RdrName (Located (body RdrName))]
Statements

-> ([Name] -> RnM (thing, FreeVars))
if these statements scope over something, this renames it and returns the result.

-> RnM (([LStmt Name (Located (body Name))], thing), FreeVars)

Rename some Stmts