RnNames

Documentation

rnImports :: [LImportDecl RdrName] -> RnM ([LImportDecl Name], GlobalRdrEnv, ImportAvails, AnyHpcUsage)

Process Import Decls Do the non SOURCE ones first, so that we get a helpful warning for SOURCE ones that are unnecessary

getLocalNonValBinders :: MiniFixityEnv -> HsGroup RdrName -> RnM ((TcGblEnv, TcLclEnv), NameSet)

rnExports :: Bool -> Maybe [LIE RdrName] -> TcGblEnv -> RnM TcGblEnv

extendGlobalRdrEnvRn :: [AvailInfo] -> MiniFixityEnv -> RnM (TcGblEnv, TcLclEnv)

gresFromAvails :: Provenance -> [AvailInfo] -> [GlobalRdrElt]

make a GlobalRdrEnv where all the elements point to the same Provenance (useful for "hiding" imports, or imports with no details).

reportUnusedNames :: Maybe [LIE RdrName] -> TcGblEnv -> RnM ()

checkConName :: RdrName -> TcRn ()