SPARC.Stack

Documentation

spRel

Arguments

:: Int
stack offset in words, positive or negative

-> AddrMode

Get an AddrMode relative to the address in sp. This gives us a stack relative addressing mode for volatile temporaries and for excess call arguments.

fpRel :: Int -> AddrMode

Get an address relative to the frame pointer. This doesn't work work for offsets greater than 13 bits; we just hope for the best

spillSlotToOffset :: Int -> Int

Convert a spill slot number to a *byte* offset, with no sign.

maxSpillSlots :: Int

The maximum number of spill slots available on the C stack. If we use up all of the slots, then we're screwed.

Why do we reserve 64 bytes, instead of using the whole thing?? -- BL 20090215