StringBuffer

Documentation

data StringBuffer

A StringBuffer is an internal pointer to a sized chunk of bytes. The bytes are intended to be *immutable*. There are pure operations to read the contents of a StringBuffer.

A StringBuffer may have a finalizer, depending on how it was obtained.

Constructors

StringBuffer

Fields

buf :: !(ForeignPtr Word8)

len :: !Int

cur :: !Int

Creation/destruction

hGetStringBuffer :: FilePath -> IO StringBuffer

hGetStringBufferBlock :: Handle -> Int -> IO StringBuffer

appendStringBuffers :: StringBuffer -> StringBuffer -> IO StringBuffer

stringToStringBuffer :: String -> StringBuffer

Inspection

nextChar :: StringBuffer -> (Char, StringBuffer)

currentChar :: StringBuffer -> Char

prevChar :: StringBuffer -> Char -> Char

atEnd :: StringBuffer -> Bool

Moving and comparison

stepOn :: StringBuffer -> StringBuffer

offsetBytes :: Int -> StringBuffer -> StringBuffer

byteDiff :: StringBuffer -> StringBuffer -> Int

Conversion

lexemeToString :: StringBuffer -> Int -> String

lexemeToFastString :: StringBuffer -> Int -> FastString

Parsing integers

parseUnsignedInteger :: StringBuffer -> Int -> Integer -> (Char -> Int) -> Integer