TargetReg

Description

Hard wired things related to registers. This is module is preventing the native code generator being able to emit code for non-host architectures.

TODO: Do a better job of the overloading, and eliminate this module. We'd probably do better with a Register type class, and hook this to Instruction somehow.

TODO: We should also make arch specific versions of RegAlloc.Graph.TrivColorable

Documentation

targetVirtualRegSqueeze :: Platform -> RegClass -> VirtualReg -> Int

targetRealRegSqueeze :: Platform -> RegClass -> RealReg -> Int

targetClassOfRealReg :: Platform -> RealReg -> RegClass

targetMkVirtualReg :: Platform -> Unique -> Format -> VirtualReg

targetRegDotColor :: Platform -> RealReg -> SDoc

targetClassOfReg :: Platform -> Reg -> RegClass