TcPluginM

Description

This module provides an interface for typechecker plugins to access select functions of the TcM, principally those to do with reading parts of the state.

Basic TcPluginM functionality

data TcPluginM a

tcPluginIO :: IO a -> TcPluginM a

Perform some IO, typically to interact with an external tool.

tcPluginTrace :: String -> SDoc -> TcPluginM ()

Output useful for debugging the compiler.

unsafeTcPluginTcM :: TcM a -> TcPluginM a

This function provides an escape for direct access to the TcM monad. It should not be used lightly, and the provided TcPluginM API should be favoured instead.

Finding Modules and Names

data FindResult

The result of searching for an imported module.

Constructors

Found ModLocation Module
The module was found

NoPackage PackageKey
The requested package was not found

FoundMultiple [(Module, ModuleOrigin)]
_Error_: both in multiple packages

NotFound
Not found

Fields

fr_paths :: [FilePath]

fr_pkg :: Maybe PackageKey

fr_mods_hidden :: [PackageKey]

fr_pkgs_hidden :: [PackageKey]

fr_suggestions :: [ModuleSuggestion]

findImportedModule :: ModuleName -> Maybe FastString -> TcPluginM FindResult

lookupOrig :: Module -> OccName -> TcPluginM Name

Looking up Names in the typechecking environment

tcLookupGlobal :: Name -> TcPluginM TyThing

tcLookupTyCon :: Name -> TcPluginM TyCon

tcLookupDataCon :: Name -> TcPluginM DataCon

tcLookupClass :: Name -> TcPluginM Class

tcLookup :: Name -> TcPluginM TcTyThing

tcLookupId :: Name -> TcPluginM Id

Getting the TcM state

getTopEnv :: TcPluginM HscEnv

getEnvs :: TcPluginM (TcGblEnv, TcLclEnv)

getInstEnvs :: TcPluginM InstEnvs

getFamInstEnvs :: TcPluginM (FamInstEnv, FamInstEnv)

matchFam :: TyCon -> [Type] -> TcPluginM (Maybe (TcCoercion, TcType))

Type variables

newFlexiTyVar :: Kind -> TcPluginM TcTyVar

isTouchableTcPluginM :: TcTyVar -> TcPluginM Bool

Zonking

zonkTcType :: TcType -> TcPluginM TcType

zonkCt :: Ct -> TcPluginM Ct