TcRnTypes

Documentation

type TcRnIf a b = IOEnv (Env a b)

type TcRn = TcRnIf TcGblEnv TcLclEnv

type TcM = TcRn

type RnM = TcRn

type IfM lcl = TcRnIf IfGblEnv lcl

type IfL = IfM IfLclEnv

type IfG = IfM ()

type TcRef a = IORef a

data Env gbl lcl

Constructors

Env

Fields

env_top :: HscEnv

env_us :: !(IORef UniqSupply)

env_gbl :: gbl

env_lcl :: lcl

data TcGblEnv

Constructors

TcGblEnv

Fields

tcg_mod :: Module

Module being compiled

tcg_src :: HscSource

What kind of module (regular Haskell, hs-boot, ext-core)

tcg_rdr_env :: GlobalRdrEnv

Top level envt; used during renaming

tcg_default :: Maybe [Type]

Types used for defaulting. Nothing => no default decl

tcg_fix_env :: FixityEnv

Just for things in this module

tcg_field_env :: RecFieldEnv

Just for things in this module See Note [The interactive package] in HscTypes

tcg_type_env :: TypeEnv

Global type env for the module we are compiling now. All TyCons and Classes (for this module) end up in here right away, along with their derived constructors, selectors.

(Ids defined in this module start in the local envt, though they move to the global envt during zonking)

NB: for what "things in this module" means, see Note [The interactive package] in HscTypes

tcg_type_env_var :: TcRef TypeEnv

tcg_inst_env :: InstEnv

Instance envt for all home-package modules; Includes the dfuns in tcg_insts

tcg_fam_inst_env :: FamInstEnv

Ditto for family instances

tcg_ann_env :: AnnEnv

And for annotations

tcg_exports :: [AvailInfo]

What is exported

tcg_imports :: ImportAvails

Information about what was imported from where, including things bound in this module. Also store Safe Haskell info here about transative trusted packaage requirements.

tcg_dus :: DefUses

What is defined in this module and what is used.

tcg_used_rdrnames :: TcRef (Set RdrName)

tcg_keep :: TcRef NameSet

Locally-defined top-level names to keep alive.

"Keep alive" means give them an Exported flag, so that the simplifier does not discard them as dead code, and so that they are exposed in the interface file (but not to export to the user).

Some things, like dict-fun Ids and default-method Ids are "born" with the Exported flag on, for exactly the above reason, but some we only discover as we go. Specifically:

The to/from functions for generic data types

Top-level variables appearing free in the RHS of an orphan rule

Top-level variables appearing free in a TH bracket

tcg_th_used :: TcRef Bool

True = Template Haskell syntax used.

We need this so that we can generate a dependency on the Template Haskell package, because the desugarer is going to emit loads of references to TH symbols. The reference is implicit rather than explicit, so we have to zap a mutable variable.

tcg_th_splice_used :: TcRef Bool

True = A Template Haskell splice was used.

Splices disable recompilation avoidance (see #481)

tcg_dfun_n :: TcRef OccSet

Allows us to choose unique DFun names.

tcg_rn_exports :: Maybe [Located (IE Name)]

tcg_rn_imports :: [LImportDecl Name]

tcg_rn_decls :: Maybe (HsGroup Name)

Renamed decls, maybe. Nothing = Don't retain renamed decls.

tcg_dependent_files :: TcRef [FilePath]

dependencies from addDependentFile

tcg_th_topdecls :: TcRef [LHsDecl RdrName]

Top-level declarations from addTopDecls

tcg_th_topnames :: TcRef NameSet

Exact names bound in top-level declarations in tcg_th_topdecls

tcg_th_modfinalizers :: TcRef [Q ()]

Template Haskell module finalizers

tcg_th_state :: TcRef (Map TypeRep Dynamic)

Template Haskell state

tcg_ev_binds :: Bag EvBind

tcg_binds :: LHsBinds Id

tcg_sigs :: NameSet

tcg_imp_specs :: [LTcSpecPrag]

tcg_warns :: Warnings

tcg_anns :: [Annotation]

tcg_tcs :: [TyCon]

tcg_insts :: [ClsInst]

tcg_fam_insts :: [FamInst]

tcg_rules :: [LRuleDecl Id]

tcg_fords :: [LForeignDecl Id]

tcg_vects :: [LVectDecl Id]

tcg_patsyns :: [PatSyn]

tcg_doc_hdr :: Maybe LHsDocString

Maybe Haddock header docs

tcg_hpc :: AnyHpcUsage

True if any part of the prog uses hpc instrumentation.

tcg_main :: Maybe Name

The Name of the main function, if this module is the main module.

tcg_safeInfer :: TcRef Bool

data TcLclEnv

Constructors

TcLclEnv

Fields

tcl_loc :: SrcSpan

tcl_ctxt :: [ErrCtxt]

tcl_untch :: Untouchables

tcl_th_ctxt :: ThStage

tcl_th_bndrs :: ThBindEnv

tcl_arrow_ctxt :: ArrowCtxt

tcl_rdr :: LocalRdrEnv

tcl_env :: TcTypeEnv

tcl_bndrs :: [TcIdBinder]

tcl_tidy :: TidyEnv

tcl_tyvars :: TcRef TcTyVarSet

tcl_lie :: TcRef WantedConstraints

tcl_errs :: TcRef Messages

data IfGblEnv

Constructors

IfGblEnv

Fields

if_rec_types :: Maybe (Module, IfG TypeEnv)

data IfLclEnv

Constructors

IfLclEnv

Fields

if_mod :: Module

if_loc :: SDoc

if_tv_env :: UniqFM TyVar

if_id_env :: UniqFM Id

type ErrCtxt = (Bool, TidyEnv -> TcM (TidyEnv, MsgDoc))

data RecFieldEnv

Constructors

RecFields (NameEnv [Name]) NameSet

data ImportAvails

ImportAvails summarises what was imported from where, irrespective of whether the imported things are actually used or not. It is used:

when processing the export list,

when constructing usage info for the interface file,

to identify the list of directly imported modules for initialisation purposes and for optimised overlap checking of family instances,

when figuring out what things are really unused

Constructors

ImportAvails

Fields

imp_mods :: ImportedMods

Domain is all directly-imported modules The ModuleName is what the module was imported as, e.g. in import Foo as Bar it is Bar.

The Bool means:

True => import was import Foo ()

False => import was some other form

Used

(a) to help construct the usage information in the interface file; if we import something we need to recompile if the export version changes

(b) to specify what child modules to initialise

We need a full ModuleEnv rather than a ModuleNameEnv here, because we might be importing modules of the same name from different packages. (currently not the case, but might be in the future).

imp_dep_mods :: ModuleNameEnv (ModuleName, IsBootInterface)

Home-package modules needed by the module being compiled

It doesn't matter whether any of these dependencies are actually used when compiling the module; they are listed if they are below it at all. For example, suppose M imports A which imports X. Then compiling M might not need to consult X.hi, but X is still listed in M's dependencies.

imp_dep_pkgs :: [PackageId]

Packages needed by the module being compiled, whether directly, or via other modules in this package, or via modules imported from other packages.

imp_trust_pkgs :: [PackageId]

This is strictly a subset of imp_dep_pkgs and records the packages the current module needs to trust for Safe Haskell compilation to succeed. A package is required to be trusted if we are dependent on a trustworthy module in that package. While perhaps making imp_dep_pkgs a tuple of (PackageId, Bool) where True for the bool indicates the package is required to be trusted is the more logical design, doing so complicates a lot of code not concerned with Safe Haskell. See Note [RnNames . Tracking Trust Transitively]

imp_trust_own_pkg :: Bool

Do we require that our own package is trusted? This is to handle efficiently the case where a Safe module imports a Trustworthy module that resides in the same package as it. See Note [RnNames . Trust Own Package]

imp_orphs :: [Module]

Orphan modules below us in the import tree (and maybe including us for imported modules)

imp_finsts :: [Module]

Family instance modules below us in the import tree (and maybe including us for imported modules)

emptyImportAvails :: ImportAvails

plusImportAvails :: ImportAvails -> ImportAvails -> ImportAvails

Union two ImportAvails

This function is a key part of Import handling, basically for each import we create a separate ImportAvails structure and then union them all together with this function.

data WhereFrom

Constructors

ImportByUser IsBootInterface

ImportBySystem

ImportByPlugin

mkModDeps :: [(ModuleName, IsBootInterface)] -> ModuleNameEnv (ModuleName, IsBootInterface)

type TcTypeEnv = NameEnv TcTyThing

data TcIdBinder

Constructors

TcIdBndr TcId TopLevelFlag

data TcTyThing

Constructors

AGlobal TyThing

ATcId

Fields

tct_id :: TcId

tct_closed :: TopLevelFlag

ATyVar Name TcTyVar

AThing TcKind

APromotionErr PromotionErr

data PromotionErr

Constructors

TyConPE

ClassPE

FamDataConPE

RecDataConPE

NoDataKinds

pprTcTyThingCategory :: TcTyThing -> SDoc

pprPECategory :: PromotionErr -> SDoc

data ThStage

Constructors

Splice Bool

Comp

Brack ThStage PendingStuff

data PendingStuff

Constructors

RnPendingUntyped (TcRef [PendingRnSplice])

RnPendingTyped

TcPending (TcRef [PendingTcSplice]) (TcRef WantedConstraints)

topStage :: ThStage

topAnnStage :: ThStage

topSpliceStage :: ThStage

type ThLevel = Int

impLevel :: ThLevel

outerLevel :: ThLevel

thLevel :: ThStage -> ThLevel

data ArrowCtxt

Constructors

NoArrowCtxt

newArrowScope :: TcM a -> TcM a

escapeArrowScope :: TcM a -> TcM a

type Xi = Type

data Ct

Constructors

CDictCan

Fields

cc_ev :: CtEvidence

cc_class :: Class

cc_tyargs :: [Xi]

CIrredEvCan

Fields

cc_ev :: CtEvidence

CTyEqCan

Fields

cc_ev :: CtEvidence

cc_tyvar :: TcTyVar

cc_rhs :: Xi

CFunEqCan

Fields

cc_ev :: CtEvidence

cc_fun :: TyCon

cc_tyargs :: [Xi]

cc_rhs :: Xi

CNonCanonical

Fields

cc_ev :: CtEvidence

CHoleCan

Fields

cc_ev :: CtEvidence

cc_occ :: OccName

type Cts = Bag Ct

emptyCts :: Cts

andCts :: Cts -> Cts -> Cts

andManyCts :: [Cts] -> Cts

dropDerivedWC :: WantedConstraints -> WantedConstraints

singleCt :: Ct -> Cts

listToCts :: [Ct] -> Cts

ctsElts :: Cts -> [Ct]

extendCts :: Cts -> Ct -> Cts

extendCtsList :: Cts -> [Ct] -> Cts

isEmptyCts :: Cts -> Bool

isCTyEqCan :: Ct -> Bool

isCFunEqCan :: Ct -> Bool

isCDictCan_Maybe :: Ct -> Maybe Class

isCFunEqCan_maybe :: Ct -> Maybe (TyCon, [Type])

isCIrredEvCan :: Ct -> Bool

isCNonCanonical :: Ct -> Bool

isWantedCt :: Ct -> Bool

isDerivedCt :: Ct -> Bool

isGivenCt :: Ct -> Bool

isHoleCt :: Ct -> Bool

ctEvidence :: Ct -> CtEvidence

ctLoc :: Ct -> CtLoc

ctPred :: Ct -> PredType

mkNonCanonical :: CtEvidence -> Ct

mkNonCanonicalCt :: Ct -> Ct

ctEvPred :: CtEvidence -> TcPredType

ctEvTerm :: CtEvidence -> EvTerm

ctEvId :: CtEvidence -> TcId

ctEvCheckDepth :: SubGoalDepth -> CtEvidence -> Bool

Checks whether the evidence can be used to solve a goal with the given minimum depth

data WantedConstraints

Constructors

WC

Fields

wc_flat :: Cts

wc_impl :: Bag Implication

wc_insol :: Cts

insolubleWC :: WantedConstraints -> Bool

emptyWC :: WantedConstraints

isEmptyWC :: WantedConstraints -> Bool

andWC :: WantedConstraints -> WantedConstraints -> WantedConstraints

unionsWC :: [WantedConstraints] -> WantedConstraints

addFlats :: WantedConstraints -> Bag Ct -> WantedConstraints

addImplics :: WantedConstraints -> Bag Implication -> WantedConstraints

mkFlatWC :: [Ct] -> WantedConstraints

addInsols :: WantedConstraints -> Bag Ct -> WantedConstraints

data Implication

Constructors

Implic

Fields

ic_untch :: Untouchables

ic_skols :: [TcTyVar]

ic_info :: SkolemInfo

ic_given :: [EvVar]

ic_fsks :: [TcTyVar]

ic_no_eqs :: Bool

ic_env :: TcLclEnv

ic_wanted :: WantedConstraints

ic_insol :: Bool

ic_binds :: EvBindsVar

data SubGoalCounter

Constructors

CountConstraints

CountTyFunApps

data SubGoalDepth

initialSubGoalDepth :: SubGoalDepth

maxSubGoalDepth :: DynFlags -> SubGoalDepth

bumpSubGoalDepth :: SubGoalCounter -> SubGoalDepth -> SubGoalDepth

subGoalCounterValue :: SubGoalCounter -> SubGoalDepth -> Int

subGoalDepthExceeded :: SubGoalDepth -> SubGoalDepth -> Maybe SubGoalCounter

data CtLoc

Constructors

CtLoc

Fields

ctl_origin :: CtOrigin

ctl_env :: TcLclEnv

ctl_depth :: !SubGoalDepth

ctLocSpan :: CtLoc -> SrcSpan

ctLocEnv :: CtLoc -> TcLclEnv

ctLocOrigin :: CtLoc -> CtOrigin

ctLocDepth :: CtLoc -> SubGoalDepth

bumpCtLocDepth :: SubGoalCounter -> CtLoc -> CtLoc

setCtLocOrigin :: CtLoc -> CtOrigin -> CtLoc

setCtLocEnv :: CtLoc -> TcLclEnv -> CtLoc

data CtOrigin

Constructors

GivenOrigin SkolemInfo

FlatSkolOrigin

OccurrenceOf Name

AppOrigin

SpecPragOrigin Name

TypeEqOrigin

Fields

uo_actual :: TcType

uo_expected :: TcType

KindEqOrigin TcType TcType CtOrigin

IPOccOrigin HsIPName

LiteralOrigin (HsOverLit Name)

NegateOrigin

ArithSeqOrigin (ArithSeqInfo Name)

PArrSeqOrigin (ArithSeqInfo Name)

SectionOrigin

TupleOrigin

AmbigOrigin UserTypeCtxt

ExprSigOrigin

PatSigOrigin

PatOrigin

RecordUpdOrigin

ViewPatOrigin

ScOrigin

DerivOrigin

DerivOriginDC DataCon Int

DerivOriginCoerce Id Type Type

StandAloneDerivOrigin

DefaultOrigin

DoOrigin

MCompOrigin

IfOrigin

ProcOrigin

AnnOrigin

FunDepOrigin

HoleOrigin

UnboundOccurrenceOf RdrName

ListOrigin

pushErrCtxt :: CtOrigin -> ErrCtxt -> CtLoc -> CtLoc

pushErrCtxtSameOrigin :: ErrCtxt -> CtLoc -> CtLoc

data SkolemInfo

Constructors

SigSkol UserTypeCtxt Type

ClsSkol Class

InstSkol

DataSkol

FamInstSkol

PatSkol ConLike (HsMatchContext Name)

ArrowSkol

IPSkol [HsIPName]

RuleSkol RuleName

InferSkol [(Name, TcType)]

BracketSkol

UnifyForAllSkol [TcTyVar] TcType

UnkSkol

data CtEvidence

Constructors

CtGiven

Fields

ctev_pred :: TcPredType

ctev_evtm :: EvTerm

ctev_loc :: CtLoc

CtWanted

Fields

ctev_pred :: TcPredType

ctev_evar :: EvVar

ctev_loc :: CtLoc

CtDerived

Fields

ctev_pred :: TcPredType

ctev_loc :: CtLoc

mkGivenLoc :: SkolemInfo -> TcLclEnv -> CtLoc

isWanted :: CtEvidence -> Bool

isGiven :: CtEvidence -> Bool

isDerived :: CtEvidence -> Bool

canRewrite :: CtEvidence -> CtEvidence -> Bool

canRewriteOrSame :: CtEvidence -> CtEvidence -> Bool

pprEvVarTheta :: [EvVar] -> SDoc

pprWantedsWithLocs :: WantedConstraints -> SDoc

pprEvVars :: [EvVar] -> SDoc

pprEvVarWithType :: EvVar -> SDoc

pprArising :: CtOrigin -> SDoc

pprArisingAt :: CtLoc -> SDoc

type TcId = Id

type TcIdSet = IdSet

data TcTyVarBind

Constructors

TcTyVarBind TcTyVar TcType

type TcTyVarBinds = Bag TcTyVarBind