TcRnTypes

Documentation

type TcRnIf a b c = IOEnv (Env a b) c

type TcRn a = TcRnIf TcGblEnv TcLclEnv a

type TcM a = TcRn a

type RnM a = TcRn a

type IfM lcl a = TcRnIf IfGblEnv lcl a

type IfL a = IfM IfLclEnv a

type IfG a = IfM () a

type TcRef a = IORef a

data Env gbl lcl

Constructors

Env

Fields

env_top :: HscEnv

env_us :: !(IORef UniqSupply)

env_gbl :: gbl

env_lcl :: lcl

data TcGblEnv

Constructors

TcGblEnv

Fields

tcg_mod :: Module

Module being compiled

tcg_src :: HscSource

What kind of module (regular Haskell, hs-boot, ext-core)

tcg_rdr_env :: GlobalRdrEnv

Top level envt; used during renaming

tcg_default :: Maybe [Type]

Types used for defaulting. Nothing => no default decl

tcg_fix_env :: FixityEnv

Just for things in this module

tcg_field_env :: RecFieldEnv

Just for things in this module

tcg_type_env :: TypeEnv

Global type env for the module we are compiling now. All TyCons and Classes (for this module) end up in here right away, along with their derived constructors, selectors.

(Ids defined in this module start in the local envt, though they move to the global envt during zonking)

tcg_type_env_var :: TcRef TypeEnv

tcg_inst_env :: InstEnv

Instance envt for home-package modules; Includes the dfuns in tcg_insts

tcg_fam_inst_env :: FamInstEnv

Ditto for family instances

tcg_exports :: [AvailInfo]

What is exported

tcg_imports :: ImportAvails

Information about what was imported from where, including things bound in this module. Also store Safe Haskell info here about transative trusted packaage requirements.

tcg_dus :: DefUses

What is defined in this module and what is used. The latter is used to generate

(a) version tracking; no need to recompile if these things have not changed version stamp

(b) unused-import info

tcg_keep :: TcRef NameSet

Locally-defined top-level names to keep alive.

Keep alive means give them an Exported flag, so that the simplifier does not discard them as dead code, and so that they are exposed in the interface file (but not to export to the user).

Some things, like dict-fun Ids and default-method Ids are born with the Exported flag on, for exactly the above reason, but some we only discover as we go. Specifically:

The to/from functions for generic data types

Top-level variables appearing free in the RHS of an orphan rule

Top-level variables appearing free in a TH bracket

tcg_th_used :: TcRef Bool

True = Template Haskell syntax used.

We need this so that we can generate a dependency on the Template Haskell package, becuase the desugarer is going to emit loads of references to TH symbols. The reference is implicit rather than explicit, so we have to zap a mutable variable.

tcg_th_splice_used :: TcRef Bool

True = A Template Haskell splice was used.

Splices disable recompilation avoidance (see #481)

tcg_dfun_n :: TcRef OccSet

Allows us to choose unique DFun names.

tcg_rn_exports :: Maybe [Located (IE Name)]

tcg_rn_imports :: [LImportDecl Name]

tcg_used_rdrnames :: TcRef (Set RdrName)

tcg_rn_decls :: Maybe (HsGroup Name)

Renamed decls, maybe. Nothing = Don't retain renamed decls.

tcg_dependent_files :: TcRef [FilePath]

dependencies from addDependentFile

tcg_ev_binds :: Bag EvBind

tcg_binds :: LHsBinds Id

tcg_sigs :: NameSet

tcg_imp_specs :: [LTcSpecPrag]

tcg_warns :: Warnings

tcg_anns :: [Annotation]

tcg_tcs :: [TyCon]

tcg_insts :: [Instance]

tcg_fam_insts :: [FamInst]

tcg_rules :: [LRuleDecl Id]

tcg_fords :: [LForeignDecl Id]

tcg_vects :: [LVectDecl Id]

tcg_doc_hdr :: Maybe LHsDocString

Maybe Haddock header docs

tcg_hpc :: AnyHpcUsage

True if any part of the prog uses hpc instrumentation.

tcg_main :: Maybe Name

The Name of the main function, if this module is the main module.

tcg_safeInfer :: TcRef Bool

data TcLclEnv

Constructors

TcLclEnv

Fields

tcl_loc :: SrcSpan

tcl_ctxt :: [ErrCtxt]

tcl_errs :: TcRef Messages

tcl_th_ctxt :: ThStage

tcl_arrow_ctxt :: ArrowCtxt

tcl_rdr :: LocalRdrEnv

tcl_env :: TcTypeEnv

tcl_tyvars :: TcRef TcTyVarSet

tcl_lie :: TcRef WantedConstraints

tcl_meta :: TcRef Unique

tcl_untch :: Unique

data IfGblEnv

Constructors

IfGblEnv

Fields

if_rec_types :: Maybe (Module, IfG TypeEnv)

data IfLclEnv

Constructors

IfLclEnv

Fields

if_mod :: Module

if_loc :: SDoc

if_tv_env :: UniqFM TyVar

if_id_env :: UniqFM Id

type ErrCtxt = (Bool, TidyEnv -> TcM (TidyEnv, Message))

data RecFieldEnv

Constructors

RecFields (NameEnv [Name]) NameSet

data ImportAvails

ImportAvails summarises what was imported from where, irrespective of whether the imported things are actually used or not. It is used:

when processing the export list,

when constructing usage info for the interface file,

to identify the list of directly imported modules for initialisation purposes and for optimised overlap checking of family instances,

when figuring out what things are really unused

Constructors

ImportAvails

Fields

imp_mods :: ImportedMods

Domain is all directly-imported modules The ModuleName is what the module was imported as, e.g. in import Foo as Bar it is Bar.

The Bool means:

True => import was import Foo ()

False => import was some other form

Used

(a) to help construct the usage information in the interface file; if we import somethign we need to recompile if the export version changes

(b) to specify what child modules to initialise

We need a full ModuleEnv rather than a ModuleNameEnv here, because we might be importing modules of the same name from different packages. (currently not the case, but might be in the future).

imp_dep_mods :: ModuleNameEnv (ModuleName, IsBootInterface)

Home-package modules needed by the module being compiled

It doesn't matter whether any of these dependencies are actually used when compiling the module; they are listed if they are below it at all. For example, suppose M imports A which imports X. Then compiling M might not need to consult X.hi, but X is still listed in M's dependencies.

imp_dep_pkgs :: [PackageId]

Packages needed by the module being compiled, whether directly, or via other modules in this package, or via modules imported from other packages.

imp_trust_pkgs :: [PackageId]

This is strictly a subset of imp_dep_pkgs and records the packages the current module needs to trust for Safe Haskell compilation to succeed. A package is required to be trusted if we are dependent on a trustworthy module in that package. While perhaps making imp_dep_pkgs a tuple of (PackageId, Bool) where True for the bool indicates the package is required to be trusted is the more logical design, doing so complicates a lot of code not concerned with Safe Haskell. See Note [RnNames . Tracking Trust Transitively]

imp_trust_own_pkg :: Bool

Do we require that our own package is trusted? This is to handle efficiently the case where a Safe module imports a Trustworthy module that resides in the same package as it. See Note [RnNames . Trust Own Package]

imp_orphs :: [Module]

Orphan modules below us in the import tree (and maybe including us for imported modules)

imp_finsts :: [Module]

Family instance modules below us in the import tree (and maybe including us for imported modules)

emptyImportAvails :: ImportAvails

plusImportAvails :: ImportAvails -> ImportAvails -> ImportAvails

Union two ImportAvails

This function is a key part of Import handling, basically for each import we create a seperate ImportAvails structure and then union them all together with this function.

data WhereFrom

Constructors

ImportByUser IsBootInterface

ImportBySystem

mkModDeps :: [(ModuleName, IsBootInterface)] -> ModuleNameEnv (ModuleName, IsBootInterface)

type TcTypeEnv = NameEnv TcTyThing

data TcTyThing

Constructors

AGlobal TyThing

ATcId

Fields

tct_id :: TcId

tct_closed :: TopLevelFlag

tct_level :: ThLevel

ATyVar Name TcType

AThing TcKind

ANothing

pprTcTyThingCategory :: TcTyThing -> SDoc

data ThStage

Constructors

Splice

Comp

Brack ThStage (TcRef [PendingSplice]) (TcRef WantedConstraints)

topStage :: ThStage

topAnnStage :: ThStage

topSpliceStage :: ThStage

type ThLevel = Int

impLevel :: ThLevel

outerLevel :: ThLevel

thLevel :: ThStage -> ThLevel

data ArrowCtxt

Constructors

NoArrowCtxt

newArrowScope :: TcM a -> TcM a

escapeArrowScope :: TcM a -> TcM a

data Untouchables

Constructors

NoUntouchables

TouchableRange Unique Unique

inTouchableRange :: Untouchables -> TcTyVar -> Bool

isNoUntouchables :: Untouchables -> Bool

type Xi = Type

data Ct

Constructors

CDictCan

Fields

cc_id :: EvVar

cc_flavor :: CtFlavor

cc_class :: Class

cc_tyargs :: [Xi]

cc_depth :: SubGoalDepth

CIPCan

Fields

cc_id :: EvVar

cc_flavor :: CtFlavor

cc_ip_nm :: IPName Name

cc_ip_ty :: TcTauType

cc_depth :: SubGoalDepth

CIrredEvCan

Fields

cc_id :: EvVar

cc_flavor :: CtFlavor

cc_ty :: Xi

cc_depth :: SubGoalDepth

CTyEqCan

Fields

cc_id :: EvVar

cc_flavor :: CtFlavor

cc_tyvar :: TcTyVar

cc_rhs :: Xi

cc_depth :: SubGoalDepth

CFunEqCan

Fields

cc_id :: EvVar

cc_flavor :: CtFlavor

cc_fun :: TyCon

cc_tyargs :: [Xi]

cc_rhs :: Xi

cc_depth :: SubGoalDepth

CNonCanonical

Fields

cc_id :: EvVar

cc_flavor :: CtFlavor

cc_depth :: SubGoalDepth

type Cts = Bag Ct

emptyCts :: Cts

andCts :: Cts -> Cts -> Cts

andManyCts :: [Cts] -> Cts

singleCt :: Ct -> Cts

extendCts :: Cts -> Ct -> Cts

isEmptyCts :: Cts -> Bool

isCTyEqCan :: Ct -> Bool

isCDictCan_Maybe :: Ct -> Maybe Class

isCIPCan_Maybe :: Ct -> Maybe (IPName Name)

isCFunEqCan_Maybe :: Ct -> Maybe TyCon

isCIrredEvCan :: Ct -> Bool

isCNonCanonical :: Ct -> Bool

type SubGoalDepth = Int

ctPred :: Ct -> PredType

data WantedConstraints

Constructors

WC

Fields

wc_flat :: Cts

wc_impl :: Bag Implication

wc_insol :: Cts

insolubleWC :: WantedConstraints -> Bool

emptyWC :: WantedConstraints

isEmptyWC :: WantedConstraints -> Bool

andWC :: WantedConstraints -> WantedConstraints -> WantedConstraints

addFlats :: WantedConstraints -> Bag WantedEvVar -> WantedConstraints

addImplics :: WantedConstraints -> Bag Implication -> WantedConstraints

mkFlatWC :: [Ct] -> WantedConstraints

data EvVarX a

Constructors

EvVarX EvVar a

mkEvVarX :: EvVar -> a -> EvVarX a

evVarOf :: EvVarX a -> EvVar

evVarX :: EvVarX a -> a

evVarOfPred :: EvVarX a -> PredType

type WantedEvVar = EvVarX WantedLoc

data Implication

Constructors

Implic

Fields

ic_untch :: Untouchables

ic_env :: TcTypeEnv

ic_skols :: TcTyVarSet

ic_given :: [EvVar]

ic_loc :: GivenLoc

ic_wanted :: WantedConstraints

ic_insol :: Bool

ic_binds :: EvBindsVar

data CtLoc orig

Constructors

CtLoc orig SrcSpan [ErrCtxt]

ctLocSpan :: CtLoc o -> SrcSpan

ctLocOrigin :: CtLoc o -> o

setCtLocOrigin :: CtLoc o -> o' -> CtLoc o'

data CtOrigin

Constructors

OccurrenceOf Name

AppOrigin

SpecPragOrigin Name

TypeEqOrigin EqOrigin

IPOccOrigin (IPName Name)

LiteralOrigin (HsOverLit Name)

NegateOrigin

ArithSeqOrigin (ArithSeqInfo Name)

PArrSeqOrigin (ArithSeqInfo Name)

SectionOrigin

TupleOrigin

AmbigOrigin Name

ExprSigOrigin

PatSigOrigin

PatOrigin

RecordUpdOrigin

ViewPatOrigin

ScOrigin

DerivOrigin

StandAloneDerivOrigin

DefaultOrigin

DoOrigin

MCompOrigin

IfOrigin

ProcOrigin

AnnOrigin

FunDepOrigin

data EqOrigin

Constructors

UnifyOrigin

Fields

uo_actual :: TcType

uo_expected :: TcType

type WantedLoc = CtLoc CtOrigin

type GivenLoc = CtLoc SkolemInfo

data GivenKind

Constructors

GivenOrig

GivenSolved (Maybe EvTerm)

pushErrCtxt :: orig -> ErrCtxt -> CtLoc orig -> CtLoc orig

data SkolemInfo

Constructors

SigSkol UserTypeCtxt Type

ClsSkol Class

InstSkol

DataSkol

FamInstSkol

PatSkol DataCon (HsMatchContext Name)

ArrowSkol

IPSkol [IPName Name]

RuleSkol RuleName

InferSkol [(Name, TcType)]

BracketSkol

UnifyForAllSkol TcType

UnkSkol

data CtFlavor

Constructors

Given GivenLoc GivenKind

Derived WantedLoc

Wanted WantedLoc

pprFlavorArising :: CtFlavor -> SDoc

isWanted :: CtFlavor -> Bool

isGivenOrSolved :: CtFlavor -> Bool

isGiven_maybe :: CtFlavor -> Maybe GivenKind

isSolved :: CtFlavor -> Bool

isDerived :: CtFlavor -> Bool

pprEvVarTheta :: [EvVar] -> SDoc

pprWantedEvVar :: WantedEvVar -> SDoc

pprWantedsWithLocs :: WantedConstraints -> SDoc

pprEvVars :: [EvVar] -> SDoc

pprEvVarWithType :: EvVar -> SDoc

pprWantedEvVarWithLoc :: WantedEvVar -> SDoc

pprArising :: CtOrigin -> SDoc

pprArisingAt :: Outputable o => CtLoc o -> SDoc

type TcId = Id

type TcIdSet = IdSet

data TcTyVarBind

Constructors

TcTyVarBind TcTyVar TcType

type TcTyVarBinds = Bag TcTyVarBind