TcSMonad

Documentation

data WorkList

Constructors

WL

Fields

wl_eqs :: [Ct]

wl_funeqs :: [Ct]

wl_rest :: [Ct]

wl_deriv :: [CtEvidence]

wl_implics :: Bag Implication

isEmptyWorkList :: WorkList -> Bool

emptyWorkList :: WorkList

extendWorkListNonEq :: Ct -> WorkList -> WorkList

extendWorkListCt :: Ct -> WorkList -> WorkList

extendWorkListDerived :: CtLoc -> CtEvidence -> WorkList -> WorkList

extendWorkListCts :: [Ct] -> WorkList -> WorkList

extendWorkListEq :: Ct -> WorkList -> WorkList

extendWorkListFunEq :: Ct -> WorkList -> WorkList

appendWorkList :: WorkList -> WorkList -> WorkList

extendWorkListImplic :: Bag Implication -> WorkList -> WorkList

selectNextWorkItem :: TcS (Maybe Ct)

workListSize :: WorkList -> Int

workListWantedCount :: WorkList -> Int

getWorkList :: TcS WorkList

updWorkListTcS :: (WorkList -> WorkList) -> TcS ()

data TcS a

runTcS :: TcS a -> TcM (a, EvBindMap)

runTcSDeriveds :: TcS a -> TcM a

This variant of runTcS will keep solving, even when only Deriveds are left around. It also doesn't return any evidence, as callers won't need it.

runTcSWithEvBinds :: EvBindsVar -> TcS a -> TcM a

failTcS :: SDoc -> TcS a

warnTcS :: WarningFlag -> SDoc -> TcS ()

addErrTcS :: SDoc -> TcS ()

runTcSEqualities :: TcS a -> TcM a

This can deal only with equality constraints.

nestTcS :: TcS a -> TcS a

nestImplicTcS :: EvBindsVar -> TcLevel -> TcS a -> TcS a

setEvBindsTcS :: EvBindsVar -> TcS a -> TcS a

buildImplication :: SkolemInfo -> [TcTyVar] -> [EvVar] -> TcS result -> TcS (Bag Implication, TcEvBinds, result)

runTcPluginTcS :: TcPluginM a -> TcS a

addUsedGRE :: Bool -> GlobalRdrElt -> TcS ()

addUsedGREs :: [GlobalRdrElt] -> TcS ()

panicTcS :: SDoc -> TcS a

traceTcS :: String -> SDoc -> TcS ()

traceFireTcS :: CtEvidence -> SDoc -> TcS ()

bumpStepCountTcS :: TcS ()

csTraceTcS :: SDoc -> TcS ()

wrapErrTcS :: TcM a -> TcS a

wrapWarnTcS :: TcM a -> TcS a

data MaybeNew

Constructors

Fresh CtEvidence

Cached EvTerm

freshGoals :: [MaybeNew] -> [CtEvidence]

isFresh :: MaybeNew -> Bool

getEvTerm :: MaybeNew -> EvTerm

newTcEvBinds :: TcS EvBindsVar

newWantedEq :: CtLoc -> Role -> TcType -> TcType -> TcS (CtEvidence, Coercion)

Make a new equality CtEvidence

emitNewWantedEq :: CtLoc -> Role -> TcType -> TcType -> TcS Coercion

Emit a new Wanted equality into the work-list

newWanted :: CtLoc -> PredType -> TcS MaybeNew

newWantedEvVar :: CtLoc -> TcPredType -> TcS MaybeNew

newWantedNC :: CtLoc -> PredType -> TcS CtEvidence

newWantedEvVarNC :: CtLoc -> TcPredType -> TcS CtEvidence

newDerivedNC :: CtLoc -> TcPredType -> TcS CtEvidence

newBoundEvVarId :: TcPredType -> EvTerm -> TcS EvVar

Make a new Id of the given type, bound (in the monad's EvBinds) to the given term

unifyTyVar :: TcTyVar -> TcType -> TcS ()

unflattenFmv :: TcTyVar -> TcType -> TcS ()

reportUnifications :: TcS a -> TcS (Int, a)

setEvBind :: EvBind -> TcS ()

setWantedEq :: TcEvDest -> Coercion -> TcS ()

Equalities only

setEqIfWanted :: CtEvidence -> Coercion -> TcS ()

Equalities only

setWantedEvTerm :: TcEvDest -> EvTerm -> TcS ()

Good for equalities and non-equalities

setWantedEvBind :: EvVar -> EvTerm -> TcS ()

setEvBindIfWanted :: CtEvidence -> EvTerm -> TcS ()

newEvVar :: TcPredType -> TcS EvVar

newGivenEvVar :: CtLoc -> (TcPredType, EvTerm) -> TcS CtEvidence

newGivenEvVars :: CtLoc -> [(TcPredType, EvTerm)] -> TcS [CtEvidence]

emitNewDerived :: CtLoc -> TcPredType -> TcS ()

emitNewDeriveds :: CtLoc -> [TcPredType] -> TcS ()

emitNewDerivedEq :: CtLoc -> Role -> TcType -> TcType -> TcS ()

checkReductionDepth

Arguments

:: CtLoc

-> TcType
type being reduced

-> TcS ()

Checks if the depth of the given location is too much. Fails if it's too big, with an appropriate error message.

getInstEnvs :: TcS InstEnvs

getFamInstEnvs :: TcS (FamInstEnv, FamInstEnv)

getTopEnv :: TcS HscEnv

getGblEnv :: TcS TcGblEnv

getLclEnv :: TcS TcLclEnv

getTcEvBindsVar :: TcS EvBindsVar

getTcLevel :: TcS TcLevel

getTcEvBindsAndTCVs :: EvBindsVar -> TcS (EvBindMap, TyCoVarSet)

getTcEvBindsMap :: TcS EvBindMap

tcLookupClass :: Name -> TcS Class

tcLookupId :: Name -> TcS Id

data InertSet

Constructors

IS

Fields

inert_cans :: InertCans

inert_flat_cache :: ExactFunEqMap (TcCoercion, TcType, CtFlavour)

inert_solved_dicts :: DictMap CtEvidence

data InertCans

Constructors

IC

Fields

inert_eqs :: InertEqs

inert_funeqs :: FunEqMap Ct

inert_dicts :: DictMap Ct

inert_safehask :: DictMap Ct

See Note [Safe Haskell Overlapping Instances Implementation] in TcSimplify

inert_irreds :: Cts

inert_insols :: Cts

inert_count :: Int

updInertTcS :: (InertSet -> InertSet) -> TcS ()

updInertCans :: (InertCans -> InertCans) -> TcS ()

updInertDicts :: (DictMap Ct -> DictMap Ct) -> TcS ()

updInertIrreds :: (Cts -> Cts) -> TcS ()

getNoGivenEqs :: TcLevel -> [TcTyVar] -> TcS (Bool, Cts)

setInertCans :: InertCans -> TcS ()

getInertEqs :: TcS (DTyVarEnv EqualCtList)

getInertCans :: TcS InertCans

getInertGivens :: TcS [Ct]

getInertInsols :: TcS Cts

emptyInert :: InertSet

getTcSInerts :: TcS InertSet

setTcSInerts :: InertSet -> TcS ()

matchableGivens :: CtLoc -> PredType -> InertSet -> Cts

Returns Given constraints that might, potentially, match the given pred. This is used when checking to see if a Given might overlap with an instance. See Note [Instance and Given overlap] in TcInteract.

prohibitedSuperClassSolve :: CtLoc -> CtLoc -> Bool

getUnsolvedInerts :: TcS (Bag Implication, Cts, Cts, Cts, Cts)

removeInertCts :: [Ct] -> InertCans -> InertCans

Remove inert constraints from the InertCans, for use when a typechecker plugin wishes to discard a given.

getPendingScDicts :: TcS [Ct]

addInertCan :: Ct -> TcS ()

addInertEq :: Ct -> TcS ()

insertFunEq :: FunEqMap a -> TyCon -> [Type] -> a -> FunEqMap a

emitInsoluble :: Ct -> TcS ()

emitWorkNC :: [CtEvidence] -> TcS ()

emitWork :: [Ct] -> TcS ()

isImprovable :: CtEvidence -> Bool

kickOutAfterUnification :: TcTyVar -> TcS Int

addInertSafehask :: InertCans -> Ct -> InertCans

insertSafeOverlapFailureTcS :: Ct -> TcS ()

updInertSafehask :: (DictMap Ct -> DictMap Ct) -> TcS ()

getSafeOverlapFailures :: TcS Cts

type DictMap a = TcAppMap a

emptyDictMap :: DictMap a

lookupInertDict :: InertCans -> Class -> [Type] -> Maybe CtEvidence

Look up a dictionary inert. NB: the returned CtEvidence might not match the input exactly. Note [Use loose types in inert set].

findDictsByClass :: DictMap a -> Class -> Bag a

addDict :: DictMap a -> Class -> [Type] -> a -> DictMap a

addDictsByClass :: DictMap Ct -> Class -> Bag Ct -> DictMap Ct

delDict :: DictMap a -> Class -> [Type] -> DictMap a

foldDicts :: (a -> b -> b) -> DictMap a -> b -> b

filterDicts :: (Ct -> Bool) -> DictMap Ct -> DictMap Ct

findDict :: DictMap a -> Class -> [Type] -> Maybe a

type EqualCtList = [Ct]

findTyEqs :: InertCans -> TyVar -> EqualCtList

foldTyEqs :: (Ct -> b -> b) -> InertEqs -> b -> b

isInInertEqs :: DTyVarEnv EqualCtList -> TcTyVar -> TcType -> Bool

lookupFlattenTyVar :: InertEqs -> TcTyVar -> TcType

lookupInertTyVar :: InertEqs -> TcTyVar -> Maybe TcType

addSolvedDict :: CtEvidence -> Class -> [Type] -> TcS ()

lookupSolvedDict :: InertSet -> Class -> [Type] -> Maybe CtEvidence

Look up a solved inert. NB: the returned CtEvidence might not match the input exactly. See Note [Use loose types in inert set].

foldIrreds :: (Ct -> b -> b) -> Cts -> b -> b

lookupFlatCache :: TyCon -> [Type] -> TcS (Maybe (TcCoercion, TcType, CtFlavour))

extendFlatCache :: TyCon -> [Type] -> (TcCoercion, TcType, CtFlavour) -> TcS ()

newFlattenSkolem :: CtFlavour -> CtLoc -> TyCon -> [TcType] -> TcS (CtEvidence, Coercion, TcTyVar)

updInertFunEqs :: (FunEqMap Ct -> FunEqMap Ct) -> TcS ()

findFunEq :: FunEqMap a -> TyCon -> [Type] -> Maybe a

findFunEqsByTyCon :: FunEqMap a -> TyCon -> [a]

instDFunType :: DFunId -> [DFunInstType] -> TcS ([TcType], TcThetaType)

newFlexiTcSTy :: Kind -> TcS TcType

instFlexi :: [TKVar] -> TcS TCvSubst

instFlexiX :: TCvSubst -> [TKVar] -> TcS TCvSubst

cloneMetaTyVar :: TcTyVar -> TcS TcTyVar

demoteUnfilledFmv :: TcTyVar -> TcS ()

tcInstType

Arguments

:: ([TyVar] -> TcM (TCvSubst, [TcTyVar]))
How to instantiate the type variables

-> Id
Type to instantiate

-> TcS ([(Name, TcTyVar)], TcThetaType, TcType)
Result (type vars, preds (incl equalities), rho)

tcInstSkolTyVarsX :: TCvSubst -> [TyVar] -> TcS (TCvSubst, [TcTyVar])

data TcLevel

isTouchableMetaTyVarTcS :: TcTyVar -> TcS Bool

isFilledMetaTyVar_maybe :: TcTyVar -> TcS (Maybe Type)

isFilledMetaTyVar :: TcTyVar -> TcS Bool

zonkTyCoVarsAndFV :: TcTyCoVarSet -> TcS TcTyCoVarSet

zonkTcType :: TcType -> TcS TcType

zonkTcTypes :: [TcType] -> TcS [TcType]

zonkTcTyVar :: TcTyVar -> TcS TcType

zonkCo :: Coercion -> TcS Coercion

zonkTyCoVarsAndFVList :: [TcTyCoVar] -> TcS [TcTyCoVar]

zonkSimples :: Cts -> TcS Cts

zonkWC :: WantedConstraints -> TcS WantedConstraints

newTcRef :: a -> TcS (TcRef a)

readTcRef :: TcRef a -> TcS a

updTcRef :: TcRef a -> (a -> a) -> TcS ()

getDefaultInfo :: TcS ([Type], (Bool, Bool))

getDynFlags :: HasDynFlags m => m DynFlags

getGlobalRdrEnvTcS :: TcS GlobalRdrEnv

matchFam :: TyCon -> [Type] -> TcS (Maybe (Coercion, TcType))

matchFamTcM :: TyCon -> [Type] -> TcM (Maybe (Coercion, TcType))

checkWellStagedDFun :: PredType -> DFunId -> CtLoc -> TcS ()

pprEq :: TcType -> TcType -> SDoc