TrieMap

Maps over Core expressions

data CoreMap a

CoreMap a is a map from CoreExpr to a. If you are a client, this is the type you want.

emptyCoreMap :: CoreMap a

extendCoreMap :: CoreMap a -> CoreExpr -> a -> CoreMap a

lookupCoreMap :: CoreMap a -> CoreExpr -> Maybe a

foldCoreMap :: (a -> b -> b) -> b -> CoreMap a -> b

Maps over Types

data TypeMap a

TypeMap a is a map from Type to a. If you are a client, this is the type you want. The keys in this map may have different kinds.

emptyTypeMap :: TypeMap a

extendTypeMap :: TypeMap a -> Type -> a -> TypeMap a

lookupTypeMap :: TypeMap a -> Type -> Maybe a

foldTypeMap :: (a -> b -> b) -> b -> TypeMap a -> b

data LooseTypeMap a

A LooseTypeMap doesn't do a kind-check. Thus, when lookup up (t |> g), you'll find entries inserted under (t), even if (g) is non-reflexive.

With explicit scoping

data CmEnv

lookupCME :: CmEnv -> Var -> Maybe BoundVar

extendTypeMapWithScope :: TypeMap a -> CmEnv -> Type -> a -> TypeMap a

Extend a TypeMap with a type in the given context. extendTypeMapWithScope m (mkDeBruijnContext [a,b,c]) t v is equivalent to extendTypeMap m (forall a b c. t) v, but allows reuse of the context over multiple insertions.

lookupTypeMapWithScope :: TypeMap a -> CmEnv -> Type -> Maybe a

mkDeBruijnContext :: [Var] -> CmEnv

Construct a deBruijn environment with the given variables in scope. e.g. mkDeBruijnEnv [a,b,c] constructs a context forall a b c.

Maps over Maybe values

data MaybeMap m a

Maps over List values

data ListMap m a

Maps over Literals

type LiteralMap a = Map Literal a

TrieMap class

class TrieMap m where

Minimal complete definition

emptyTM, lookupTM, alterTM, mapTM, foldTM

Associated Types

type Key m :: *

Methods

emptyTM :: m a

lookupTM :: forall b. Key m -> m b -> Maybe b

alterTM :: forall b. Key m -> XT b -> m b -> m b

mapTM :: (a -> b) -> m a -> m b

foldTM :: (a -> b -> b) -> m a -> b -> b

insertTM :: TrieMap m => Key m -> a -> m a -> m a

deleteTM :: TrieMap m => Key m -> m a -> m a

lkDFreeVar :: Var -> DVarEnv a -> Maybe a

xtDFreeVar :: Var -> XT a -> DVarEnv a -> DVarEnv a

lkDNamed :: NamedThing n => n -> DNameEnv a -> Maybe a

xtDNamed :: NamedThing n => n -> XT a -> DNameEnv a -> DNameEnv a

(>.>) :: (a -> b) -> (b -> c) -> a -> c infixr 1

(|>) :: a -> (a -> b) -> b infixr 1

(|>>) :: TrieMap m2 => (XT (m2 a) -> m1 (m2 a) -> m1 (m2 a)) -> (m2 a -> m2 a) -> m1 (m2 a) -> m1 (m2 a) infixr 1