TrieMap

Documentation

data CoreMap a

CoreMap a is a map from CoreExpr to a. If you are a client, this is the type you want.

emptyCoreMap :: CoreMap a

extendCoreMap :: CoreMap a -> CoreExpr -> a -> CoreMap a

lookupCoreMap :: CoreMap a -> CoreExpr -> Maybe a

foldCoreMap :: (a -> b -> b) -> b -> CoreMap a -> b

data TypeMap a

TypeMap a is a map from Type to a. If you are a client, this is the type you want. The keys in this map may have different kinds.

emptyTypeMap :: TypeMap a

extendTypeMap :: TypeMap a -> Type -> a -> TypeMap a

lookupTypeMap :: TypeMap a -> Type -> Maybe a

foldTypeMap :: (a -> b -> b) -> b -> TypeMap a -> b

data LooseTypeMap a

A LooseTypeMap doesn't do a kind-check. Thus, when lookup up (t |> g), you'll find entries inserted under (t), even if (g) is non-reflexive.

data MaybeMap m a

data ListMap m a

class TrieMap m where

Minimal complete definition

emptyTM, lookupTM, alterTM, mapTM, foldTM

Associated Types

type Key m :: *

Methods

emptyTM :: m a

lookupTM :: forall b. Key m -> m b -> Maybe b

alterTM :: forall b. Key m -> XT b -> m b -> m b

mapTM :: (a -> b) -> m a -> m b

foldTM :: (a -> b -> b) -> m a -> b -> b

insertTM :: TrieMap m => Key m -> a -> m a -> m a

deleteTM :: TrieMap m => Key m -> m a -> m a