TyCoRep

Documentation

data TyThing

A global typecheckable-thing, essentially anything that has a name. Not to be confused with a TcTyThing, which is also a typecheckable thing but in the *local* context. See TcEnv for how to retrieve a TyThing given a Name.

Constructors

AnId Id

AConLike ConLike

ATyCon TyCon

ACoAxiom (CoAxiom Branched)

tyThingCategory :: TyThing -> String

pprTyThingCategory :: TyThing -> SDoc

pprShortTyThing :: TyThing -> SDoc

Types

data Type

Constructors

TyVarTy Var
Vanilla type or kind variable (*never* a coercion variable)

AppTy Type Type
Type application to something other than a TyCon. Parameters:

1) Function: must not be a TyConApp or CastTy, must be another AppTy, or TyVarTy See Note Respecting definitional equality about the no CastTy requirement

2) Argument type

TyConApp TyCon [KindOrType]
Application of a TyCon, including newtypes and synonyms. Invariant: saturated applications of FunTyCon must use FunTy and saturated synonyms must use their own constructors. However, unsaturated FunTyCons do appear as TyConApps. Parameters:

1) Type constructor being applied to.

2) Type arguments. Might not have enough type arguments here to saturate the constructor. Even type synonyms are not necessarily saturated; for example unsaturated type synonyms can appear as the right hand side of a type synonym.

ForAllTy !TyCoVarBinder Type
A Π type.

FunTy Type Type
t1 -> t2 Very common, so an important special case

LitTy TyLit
Type literals are similar to type constructors.

CastTy Type KindCoercion
A kind cast. The coercion is always nominal. INVARIANT: The cast is never refl. INVARIANT: The Type is not a CastTy (use TransCo instead) See Note Respecting definitional equality and (EQ3)

CoercionTy Coercion
Injection of a Coercion into a type This should only ever be used in the RHS of an AppTy, in the list of a TyConApp, when applying a promoted GADT data constructor

data TyLit

Constructors

NumTyLit Integer

StrTyLit FastString

type KindOrType = Type

The key representation of types within the compiler

type Kind = Type

The key type representing kinds in the compiler.

type KnotTied ty = ty

A type labeled KnotTied might have knot-tied tycons in it. See Note [Type checking recursive type and class declarations] in TcTyClsDecls

type PredType = Type

A type of the form p of kind Constraint represents a value whose type is the Haskell predicate p, where a predicate is what occurs before the => in a Haskell type.

We use PredType as documentation to mark those types that we guarantee to have this kind.

It can be expanded into its representation, but:

The type checker must treat it as opaque

The rest of the compiler treats it as transparent

Consider these examples:

f :: (Eq a) => a -> Int
g :: (?x :: Int -> Int) => a -> Int
h :: (r\l) => {r} => {l::Int | r}

Here the Eq a and ?x :: Int -> Int and rl are all called "predicates"

type ThetaType = [PredType]

A collection of PredTypes

data ArgFlag

Argument Flag

Is something required to appear in source Haskell (Required), permitted by request (Specified) (visible type application), or prohibited entirely from appearing in source Haskell (Inferred)? See Note [VarBndrs, TyCoVarBinders, TyConBinders, and visibility] in TyCoRep

Constructors

Inferred

Specified

Required

Coercions

data Coercion

A Coercion is concrete evidence of the equality/convertibility of two types.

Constructors

Refl Type

GRefl Role Type MCoercionN

TyConAppCo Role TyCon [Coercion]

AppCo Coercion CoercionN

ForAllCo TyCoVar KindCoercion Coercion

FunCo Role Coercion Coercion

CoVarCo CoVar

AxiomInstCo (CoAxiom Branched) BranchIndex [Coercion]

AxiomRuleCo CoAxiomRule [Coercion]

UnivCo UnivCoProvenance Role Type Type

SymCo Coercion

TransCo Coercion Coercion

NthCo Role Int Coercion

LRCo LeftOrRight CoercionN

InstCo Coercion CoercionN

KindCo Coercion

SubCo CoercionN

HoleCo CoercionHole
See Note [Coercion holes] Only present during typechecking

data UnivCoProvenance

For simplicity, we have just one UnivCo that represents a coercion from some type to some other type, with (in general) no restrictions on the type. The UnivCoProvenance specifies more exactly what the coercion really is and why a program should (or shouldn't!) trust the coercion. It is reasonable to consider each constructor of UnivCoProvenance as a totally independent coercion form; their only commonality is that they don't tell you what types they coercion between. (That info is in the UnivCo constructor of Coercion.

Constructors

UnsafeCoerceProv
From unsafeCoerce#. These are unsound.

PhantomProv KindCoercion
See Note [Phantom coercions]. Only in Phantom roled coercions

ProofIrrelProv KindCoercion
From the fact that any two coercions are considered equivalent. See Note [ProofIrrelProv]. Can be used in Nominal or Representational coercions

PluginProv String
From a plugin, which asserts that this coercion is sound. The string is for the use of the plugin.

data CoercionHole

A coercion to be filled in by the type-checker. See Note [Coercion holes]

Constructors

CoercionHole

Fields

ch_co_var :: CoVar

ch_ref :: IORef (Maybe Coercion)

coHoleCoVar :: CoercionHole -> CoVar

setCoHoleCoVar :: CoercionHole -> CoVar -> CoercionHole

type CoercionN = Coercion

type CoercionR = Coercion

type CoercionP = Coercion

type KindCoercion = CoercionN

data MCoercion

A semantically more meaningful type to represent what may or may not be a useful Coercion.

Constructors

MRefl

MCo Coercion

type MCoercionR = MCoercion

type MCoercionN = MCoercion

Functions over types

mkTyConTy :: TyCon -> Type

Create the plain type constructor type which has been applied to no type arguments at all.

mkTyVarTy :: TyVar -> Type

mkTyVarTys :: [TyVar] -> [Type]

mkTyCoVarTy :: TyCoVar -> Type

mkTyCoVarTys :: [TyCoVar] -> [Type]

mkFunTy :: Type -> Type -> Type infixr 3

Make an arrow type

mkFunTys :: [Type] -> Type -> Type

Make nested arrow types

mkTyCoForAllTy :: TyCoVar -> ArgFlag -> Type -> Type

If tv is a coercion variable and it is not used in the body, returns a FunTy, otherwise makes a forall type. See Note [Unused coercion variable in ForAllTy]

mkForAllTys :: [TyCoVarBinder] -> Type -> Type

Wraps foralls over the type using the provided TyCoVars from left to right

mkForAllTy :: TyCoVar -> ArgFlag -> Type -> Type

Like mkTyCoForAllTy, but does not check the occurrence of the binder See Note [Unused coercion variable in ForAllTy]

mkTyCoPiTy :: TyCoBinder -> Type -> Type

mkTyCoPiTys :: [TyCoBinder] -> Type -> Type

mkPiTys :: [TyCoBinder] -> Type -> Type

Like mkTyCoPiTys, but does not check the occurrence of the binder

kindRep_maybe :: HasDebugCallStack => Kind -> Maybe Type

Given a kind (TYPE rr), extract its RuntimeRep classifier rr. For example, kindRep_maybe * = Just LiftedRep Returns Nothing if the kind is not of form (TYPE rr) Treats * and Constraint as the same

kindRep :: HasDebugCallStack => Kind -> Type

Extract the RuntimeRep classifier of a type from its kind. For example, kindRep * = LiftedRep; Panics if this is not possible. Treats * and Constraint as the same

isLiftedTypeKind :: Kind -> Bool

This version considers Constraint to be the same as *. Returns True if the argument is equivalent to Type/Constraint and False otherwise. See Note [Kind Constraint and kind Type]

isUnliftedTypeKind :: Kind -> Bool

Returns True if the kind classifies unlifted types and False otherwise. Note that this returns False for levity-polymorphic kinds, which may be specialized to a kind that classifies unlifted types.

isLiftedRuntimeRep :: Type -> Bool

isUnliftedRuntimeRep :: Type -> Bool

isRuntimeRepTy :: Type -> Bool

Is this the type RuntimeRep?

isRuntimeRepVar :: TyVar -> Bool

Is a tyvar of type RuntimeRep?

sameVis :: ArgFlag -> ArgFlag -> Bool

Do these denote the same level of visibility? Required arguments are visible, others are not. So this function equates Specified and Inferred. Used for printing.

Functions over binders

data TyCoBinder

A TyCoBinder represents an argument to a function. TyCoBinders can be dependent (Named) or nondependent (Anon). They may also be visible or not. See Note [TyCoBinders]

Constructors

Named TyCoVarBinder

Anon Type

type TyCoVarBinder = VarBndr TyCoVar ArgFlag

Variable Binder

A TyCoVarBinder is the binder of a ForAllTy It's convenient to define this synonym here rather its natural home in TyCoRep, because it's used in DataCon.hs-boot

A TyVarBinder is a binder with only TyVar

type TyBinder = TyCoBinder

TyBinder is like TyCoBinder, but there can only be TyVarBinder in the Named field.

binderVar :: VarBndr tv argf -> tv

binderVars :: [VarBndr tv argf] -> [tv]

binderType :: VarBndr TyCoVar argf -> Type

binderArgFlag :: VarBndr tv argf -> argf

delBinderVar :: VarSet -> TyCoVarBinder -> VarSet

Remove the binder's variable from the set, if the binder has a variable.

isInvisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is not written in Haskell?

isVisibleArgFlag :: ArgFlag -> Bool

Does this ArgFlag classify an argument that is written in Haskell?

isInvisibleBinder :: TyCoBinder -> Bool

Does this binder bind an invisible argument?

isVisibleBinder :: TyCoBinder -> Bool

Does this binder bind a visible argument?

isTyBinder :: TyCoBinder -> Bool

If its a named binder, is the binder a tyvar? Returns True for nondependent binder.

isNamedBinder :: TyCoBinder -> Bool

tyCoBinderArgFlag :: TyCoBinder -> ArgFlag

Functions over coercions

pickLR :: LeftOrRight -> (a, a) -> a

Pretty-printing

pprType :: Type -> SDoc

pprParendType :: Type -> SDoc

pprPrecType :: PprPrec -> Type -> SDoc

pprPrecTypeX :: TidyEnv -> PprPrec -> Type -> SDoc

pprTypeApp :: TyCon -> [Type] -> SDoc

pprTCvBndr :: TyCoVarBinder -> SDoc

pprTCvBndrs :: [TyCoVarBinder] -> SDoc

pprSigmaType :: Type -> SDoc

pprTheta :: ThetaType -> SDoc

pprParendTheta :: ThetaType -> SDoc

pprForAll :: [TyCoVarBinder] -> SDoc

pprUserForAll :: [TyCoVarBinder] -> SDoc

Print a user-level forall; see Note [When to print foralls]

pprTyVar :: TyVar -> SDoc

pprTyVars :: [TyVar] -> SDoc

pprThetaArrowTy :: ThetaType -> SDoc

pprClassPred :: Class -> [Type] -> SDoc

pprKind :: Kind -> SDoc

pprParendKind :: Kind -> SDoc

pprTyLit :: TyLit -> SDoc

newtype PprPrec

A general-purpose pretty-printing precedence type.

Constructors

PprPrec Int

topPrec :: PprPrec

sigPrec :: PprPrec

opPrec :: PprPrec

funPrec :: PprPrec

appPrec :: PprPrec

maybeParen :: PprPrec -> PprPrec -> SDoc -> SDoc

pprDataCons :: TyCon -> SDoc

pprWithExplicitKindsWhen :: Bool -> SDoc -> SDoc

Display all kind information (with -fprint-explicit-kinds) when the provided Bool argument is True. See Note [Kind arguments in error messages] in TcErrors.

pprCo :: Coercion -> SDoc

pprParendCo :: Coercion -> SDoc

debugPprType :: Type -> SDoc

debugPprType is a simple pretty printer that prints a type without going through IfaceType. It does not format as prettily as the normal route, but it's much more direct, and that can be useful for debugging. E.g. with -dppr-debug it prints the kind on type-variable occurrences which the normal route fundamentally cannot do.

Free variables

tyCoVarsOfType :: Type -> TyCoVarSet

tyCoVarsOfTypeDSet :: Type -> DTyCoVarSet

tyCoFVsOfType that returns free variables of a type in a deterministic set. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in FV.

tyCoVarsOfTypes :: [Type] -> TyCoVarSet

tyCoVarsOfTypesDSet :: [Type] -> DTyCoVarSet

Returns free variables of types, including kind variables as a deterministic set. For type synonyms it does not expand the synonym.

tyCoFVsBndr :: TyCoVarBinder -> FV -> FV

tyCoFVsVarBndr :: Var -> FV -> FV

tyCoFVsVarBndrs :: [Var] -> FV -> FV

tyCoFVsOfType :: Type -> FV

The worker for tyCoFVsOfType and tyCoFVsOfTypeList. The previous implementation used unionVarSet which is O(n+m) and can make the function quadratic. It's exported, so that it can be composed with other functions that compute free variables. See Note [FV naming conventions] in FV.

Eta-expanded because that makes it run faster (apparently) See Note [FV eta expansion] in FV for explanation.

tyCoVarsOfTypeList :: Type -> [TyCoVar]

tyCoFVsOfType that returns free variables of a type in deterministic order. For explanation of why using VarSet is not deterministic see Note [Deterministic FV] in FV.

tyCoFVsOfTypes :: [Type] -> FV

tyCoVarsOfTypesList :: [Type] -> [TyCoVar]

Returns free variables of types, including kind variables as a deterministically ordered list. For type synonyms it does not expand the synonym.

coVarsOfType :: Type -> CoVarSet

coVarsOfTypes :: [Type] -> TyCoVarSet

coVarsOfCo :: Coercion -> CoVarSet

coVarsOfCos :: [Coercion] -> CoVarSet

tyCoVarsOfCo :: Coercion -> TyCoVarSet

tyCoVarsOfCos :: [Coercion] -> TyCoVarSet

tyCoVarsOfCoDSet :: Coercion -> DTyCoVarSet

Get a deterministic set of the vars free in a coercion

tyCoFVsOfCo :: Coercion -> FV

tyCoFVsOfCos :: [Coercion] -> FV

tyCoVarsOfCoList :: Coercion -> [TyCoVar]

tyCoVarsOfProv :: UnivCoProvenance -> TyCoVarSet

almostDevoidCoVarOfCo :: CoVar -> Coercion -> Bool

Given a covar and a coercion, returns True if covar is almost devoid in the coercion. That is, covar can only appear in Refl and GRefl. See last wrinkle in Note [Unused coercion variable in ForAllCo] in Coercion

injectiveVarsOfType :: Type -> FV

Returns the free variables of a Type that are in injective positions. For example, if F is a non-injective type family, then:

injectiveTyVarsOf( Either c (Maybe (a, F b c)) ) = {a,c}

If injectiveVarsOfType ty = itvs, then knowing ty fixes itvs. More formally, if a is in injectiveVarsOfType ty and S1(ty) ~ S2(ty), then S1(a) ~ S2(a), where S1 and S2 are arbitrary substitutions.

See Note [When does a tycon application need an explicit kind signature?].

tyConAppNeedsKindSig

Arguments

:: Bool
Should specified binders count towards injective positions in the kind of the TyCon?

-> TyCon

-> Int
The number of args the TyCon is applied to.

-> Bool
Does T t_1 ... t_n need a kind signature? (Where n is the number of arguments)

Does a TyCon (that is applied to some number of arguments) need to be ascribed with an explicit kind signature to resolve ambiguity if rendered as a source-syntax type? (See Note [When does a tycon application need an explicit kind signature?] for a full explanation of what this function checks for.)

noFreeVarsOfType :: Type -> Bool

Returns True if this type has no free variables. Should be the same as isEmptyVarSet . tyCoVarsOfType, but faster in the non-forall case.

noFreeVarsOfCo :: Coercion -> Bool

Returns True if this coercion has no free variables. Should be the same as isEmptyVarSet . tyCoVarsOfCo, but faster in the non-forall case.

Substitutions

data TCvSubst

Type & coercion substitution

The following invariants must hold of a TCvSubst:

The in-scope set is needed only to guide the generation of fresh uniques

In particular, the kind of the type variables in the in-scope set is not relevant

The substitution is only applied ONCE! This is because in general such application will not reach a fixed point.

Constructors

TCvSubst InScopeSet TvSubstEnv CvSubstEnv

type TvSubstEnv = TyVarEnv Type

A substitution of Types for TyVars and Kinds for KindVars

type CvSubstEnv = CoVarEnv Coercion

A substitution of Coercions for CoVars

emptyTvSubstEnv :: TvSubstEnv

emptyCvSubstEnv :: CvSubstEnv

composeTCvSubstEnv :: InScopeSet -> (TvSubstEnv, CvSubstEnv) -> (TvSubstEnv, CvSubstEnv) -> (TvSubstEnv, CvSubstEnv)

(compose env1 env2)(x) is env1(env2(x)); i.e. apply env2 then env1. It assumes that both are idempotent. Typically, env1 is the refinement to a base substitution env2

composeTCvSubst :: TCvSubst -> TCvSubst -> TCvSubst

Composes two substitutions, applying the second one provided first, like in function composition.

emptyTCvSubst :: TCvSubst

mkEmptyTCvSubst :: InScopeSet -> TCvSubst

isEmptyTCvSubst :: TCvSubst -> Bool

mkTCvSubst :: InScopeSet -> (TvSubstEnv, CvSubstEnv) -> TCvSubst

mkTvSubst :: InScopeSet -> TvSubstEnv -> TCvSubst

Make a TCvSubst with specified tyvar subst and empty covar subst

mkCvSubst :: InScopeSet -> CvSubstEnv -> TCvSubst

Make a TCvSubst with specified covar subst and empty tyvar subst

getTvSubstEnv :: TCvSubst -> TvSubstEnv

getCvSubstEnv :: TCvSubst -> CvSubstEnv

getTCvInScope :: TCvSubst -> InScopeSet

getTCvSubstRangeFVs :: TCvSubst -> VarSet

Returns the free variables of the types in the range of a substitution as a non-deterministic set.

isInScope :: Var -> TCvSubst -> Bool

notElemTCvSubst :: Var -> TCvSubst -> Bool

setTvSubstEnv :: TCvSubst -> TvSubstEnv -> TCvSubst

setCvSubstEnv :: TCvSubst -> CvSubstEnv -> TCvSubst

zapTCvSubst :: TCvSubst -> TCvSubst

extendTCvInScope :: TCvSubst -> Var -> TCvSubst

extendTCvInScopeList :: TCvSubst -> [Var] -> TCvSubst

extendTCvInScopeSet :: TCvSubst -> VarSet -> TCvSubst

extendTCvSubst :: TCvSubst -> TyCoVar -> Type -> TCvSubst

extendTCvSubstWithClone :: TCvSubst -> TyCoVar -> TyCoVar -> TCvSubst

extendCvSubst :: TCvSubst -> CoVar -> Coercion -> TCvSubst

extendCvSubstWithClone :: TCvSubst -> CoVar -> CoVar -> TCvSubst

extendTvSubst :: TCvSubst -> TyVar -> Type -> TCvSubst

extendTvSubstBinderAndInScope :: TCvSubst -> TyCoBinder -> Type -> TCvSubst

extendTvSubstWithClone :: TCvSubst -> TyVar -> TyVar -> TCvSubst

extendTvSubstList :: TCvSubst -> [Var] -> [Type] -> TCvSubst

extendTvSubstAndInScope :: TCvSubst -> TyVar -> Type -> TCvSubst

extendTCvSubstList :: TCvSubst -> [Var] -> [Type] -> TCvSubst

unionTCvSubst :: TCvSubst -> TCvSubst -> TCvSubst

zipTyEnv :: HasDebugCallStack => [TyVar] -> [Type] -> TvSubstEnv

zipCoEnv :: HasDebugCallStack => [CoVar] -> [Coercion] -> CvSubstEnv

mkTyCoInScopeSet :: [Type] -> [Coercion] -> InScopeSet

Generates an in-scope set from the free variables in a list of types and a list of coercions

zipTvSubst :: HasDebugCallStack => [TyVar] -> [Type] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please!

zipCvSubst :: HasDebugCallStack => [CoVar] -> [Coercion] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No TyVars, please!

zipTCvSubst :: HasDebugCallStack => [TyCoVar] -> [Type] -> TCvSubst

mkTvSubstPrs :: [(TyVar, Type)] -> TCvSubst

Generates the in-scope set for the TCvSubst from the types in the incoming environment. No CoVars, please!

substTyWith :: HasCallStack => [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst

substTyWithCoVars :: [CoVar] -> [Coercion] -> Type -> Type

Substitute covars within a type

substTysWith :: [TyVar] -> [Type] -> [Type] -> [Type]

Type substitution, see zipTvSubst

substTysWithCoVars :: [CoVar] -> [Coercion] -> [Type] -> [Type]

Type substitution, see zipTvSubst

substCoWith :: HasCallStack => [TyVar] -> [Type] -> Coercion -> Coercion

Coercion substitution, see zipTvSubst

substTy :: HasCallStack => TCvSubst -> Type -> Type

Substitute within a Type The substitution has to satisfy the invariants described in Note [The substitution invariant].

substTyAddInScope :: TCvSubst -> Type -> Type

Substitute within a Type after adding the free variables of the type to the in-scope set. This is useful for the case when the free variables aren't already in the in-scope set or easily available. See also Note [The substitution invariant].

substTyUnchecked :: TCvSubst -> Type -> Type

Substitute within a Type disabling the sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substTysUnchecked :: TCvSubst -> [Type] -> [Type]

Substitute within several Types disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTysUnchecked to substTys and remove this function. Please don't use in new code.

substThetaUnchecked :: TCvSubst -> ThetaType -> ThetaType

Substitute within a ThetaType disabling the sanity checks. The problems that the sanity checks in substTys catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substThetaUnchecked to substTheta and remove this function. Please don't use in new code.

substTyWithUnchecked :: [TyVar] -> [Type] -> Type -> Type

Type substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substTy catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substTyUnchecked to substTy and remove this function. Please don't use in new code.

substCoUnchecked :: TCvSubst -> Coercion -> Coercion

Substitute within a Coercion disabling sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substCoWithUnchecked :: [TyVar] -> [Type] -> Coercion -> Coercion

Coercion substitution, see zipTvSubst. Disables sanity checks. The problems that the sanity checks in substCo catch are described in Note [The substitution invariant]. The goal of #11371 is to migrate all the calls of substCoUnchecked to substCo and remove this function. Please don't use in new code.

substTyWithInScope :: InScopeSet -> [TyVar] -> [Type] -> Type -> Type

Substitute tyvars within a type using a known InScopeSet. Pre-condition: the in_scope set should satisfy Note [The substitution invariant]; specifically it should include the free vars of tys, and of ty minus the domain of the subst.

substTys :: HasCallStack => TCvSubst -> [Type] -> [Type]

Substitute within several Types The substitution has to satisfy the invariants described in Note [The substitution invariant].

substTheta :: HasCallStack => TCvSubst -> ThetaType -> ThetaType

Substitute within a ThetaType The substitution has to satisfy the invariants described in Note [The substitution invariant].

lookupTyVar :: TCvSubst -> TyVar -> Maybe Type

substCo :: HasCallStack => TCvSubst -> Coercion -> Coercion

Substitute within a Coercion The substitution has to satisfy the invariants described in Note [The substitution invariant].

substCos :: HasCallStack => TCvSubst -> [Coercion] -> [Coercion]

Substitute within several Coercions The substitution has to satisfy the invariants described in Note [The substitution invariant].

substCoVar :: TCvSubst -> CoVar -> Coercion

substCoVars :: TCvSubst -> [CoVar] -> [Coercion]

lookupCoVar :: TCvSubst -> Var -> Maybe Coercion

cloneTyVarBndr :: TCvSubst -> TyVar -> Unique -> (TCvSubst, TyVar)

cloneTyVarBndrs :: TCvSubst -> [TyVar] -> UniqSupply -> (TCvSubst, [TyVar])

substVarBndr :: HasCallStack => TCvSubst -> TyCoVar -> (TCvSubst, TyCoVar)

substVarBndrs :: HasCallStack => TCvSubst -> [TyCoVar] -> (TCvSubst, [TyCoVar])

substTyVarBndr :: HasCallStack => TCvSubst -> TyVar -> (TCvSubst, TyVar)

substTyVarBndrs :: HasCallStack => TCvSubst -> [TyVar] -> (TCvSubst, [TyVar])

substCoVarBndr :: HasCallStack => TCvSubst -> CoVar -> (TCvSubst, CoVar)

substTyVar :: TCvSubst -> TyVar -> Type

substTyVars :: TCvSubst -> [TyVar] -> [Type]

substTyCoVars :: TCvSubst -> [TyCoVar] -> [Type]

substForAllCoBndr :: TCvSubst -> TyCoVar -> KindCoercion -> (TCvSubst, TyCoVar, Coercion)

substVarBndrUsing :: (TCvSubst -> Type -> Type) -> TCvSubst -> TyCoVar -> (TCvSubst, TyCoVar)

substForAllCoBndrUsing :: Bool -> (Coercion -> Coercion) -> TCvSubst -> TyCoVar -> KindCoercion -> (TCvSubst, TyCoVar, KindCoercion)

checkValidSubst :: HasCallStack => TCvSubst -> [Type] -> [Coercion] -> a -> a

This checks if the substitution satisfies the invariant from Note [The substitution invariant].

isValidTCvSubst :: TCvSubst -> Bool

When calling substTy it should be the case that the in-scope set in the substitution is a superset of the free vars of the range of the substitution. See also Note [The substitution invariant].

Tidying type related things up for printing

tidyType :: TidyEnv -> Type -> Type

tidyTypes :: TidyEnv -> [Type] -> [Type]

tidyOpenType :: TidyEnv -> Type -> (TidyEnv, Type)

tidyOpenTypes :: TidyEnv -> [Type] -> (TidyEnv, [Type])

Grabs the free type variables, tidies them and then uses tidyType to work over the type itself

tidyOpenKind :: TidyEnv -> Kind -> (TidyEnv, Kind)

tidyVarBndr :: TidyEnv -> TyCoVar -> (TidyEnv, TyCoVar)

tidyVarBndrs :: TidyEnv -> [TyCoVar] -> (TidyEnv, [TyCoVar])

This tidies up a type for printing in an error message, or in an interface file.

It doesn't change the uniques at all, just the print names.

tidyFreeTyCoVars :: TidyEnv -> [TyCoVar] -> TidyEnv

Add the free TyVars to the env in tidy form, so that we can tidy the type they are free in

avoidNameClashes :: [TyCoVar] -> TidyEnv -> TidyEnv

tidyOpenTyCoVar :: TidyEnv -> TyCoVar -> (TidyEnv, TyCoVar)

Treat a new TyCoVar as a binder, and give it a fresh tidy name using the environment if one has not already been allocated. See also tidyVarBndr

tidyOpenTyCoVars :: TidyEnv -> [TyCoVar] -> (TidyEnv, [TyCoVar])

tidyTyCoVarOcc :: TidyEnv -> TyCoVar -> TyCoVar

tidyTopType :: Type -> Type

Calls tidyType on a top-level type (i.e. with an empty tidying environment)

tidyKind :: TidyEnv -> Kind -> Kind

tidyCo :: TidyEnv -> Coercion -> Coercion

tidyCos :: TidyEnv -> [Coercion] -> [Coercion]

tidyTyCoVarBinder :: TidyEnv -> VarBndr TyCoVar vis -> (TidyEnv, VarBndr TyCoVar vis)

tidyTyCoVarBinders :: TidyEnv -> [VarBndr TyCoVar vis] -> (TidyEnv, [VarBndr TyCoVar vis])

Sizes

typeSize :: Type -> Int

coercionSize :: Coercion -> Int

provSize :: UnivCoProvenance -> Int