TysWiredIn

Description

This module is about types that can be defined in Haskell, but which must be wired into the compiler nonetheless. C.f module TysPrim

All wired in things

wiredInTyCons :: [TyCon]

Bool

boolTy :: Type

boolTyCon :: TyCon

boolTyCon_RDR :: RdrName

boolTyConName :: Name

trueDataCon :: DataCon

trueDataConId :: Id

true_RDR :: RdrName

falseDataCon :: DataCon

falseDataConId :: Id

false_RDR :: RdrName

Ordering

ltDataCon :: DataCon

ltDataConId :: Id

eqDataCon :: DataCon

eqDataConId :: Id

gtDataCon :: DataCon

gtDataConId :: Id

Char

charTyCon :: TyCon

charDataCon :: DataCon

charTyCon_RDR :: RdrName

charTy :: Type

stringTy :: Type

charTyConName :: Name

integerGmpSDataCon :: DataCon

Double

doubleTyCon :: TyCon

doubleDataCon :: DataCon

doubleTy :: Type

doubleTyConName :: Name

Float

floatTyCon :: TyCon

floatDataCon :: DataCon

floatTy :: Type

floatTyConName :: Name

Int

intTyCon :: TyCon

intDataCon :: DataCon

intTyCon_RDR :: RdrName

intDataCon_RDR :: RdrName

intTyConName :: Name

intTy :: Type

Word

wordTyCon :: TyCon

wordDataCon :: DataCon

wordTyConName :: Name

wordTy :: Type

List

listTyCon :: TyCon

nilDataCon :: DataCon

consDataCon :: DataCon

listTyCon_RDR :: RdrName

consDataCon_RDR :: RdrName

listTyConName :: Name

mkListTy :: Type -> Type

Tuples

mkTupleTy :: TupleSort -> [Type] -> Type

mkBoxedTupleTy :: [Type] -> Type

Build the type of a small tuple that holds the specified type of thing

tupleTyCon :: TupleSort -> Arity -> TyCon

tupleCon :: TupleSort -> Arity -> DataCon

unitTyCon :: TyCon

unitDataCon :: DataCon

unitDataConId :: Id

pairTyCon :: TyCon

unboxedSingletonTyCon :: TyCon

unboxedSingletonDataCon :: DataCon

unboxedPairTyCon :: TyCon

unboxedPairDataCon :: DataCon

Unit

unitTy :: Type

Parallel arrays

mkPArrTy :: Type -> Type

Construct a type representing the application of the parallel array constructor

parrTyCon :: TyCon

Represents the type constructor of parallel arrays

This must match the definition in PrelPArr

NB: Although the constructor is given here, it will not be accessible in user code as it is not in the environment of any compiled module except PrelPArr.

parrFakeCon :: Arity -> DataCon

Fake array constructors

These constructors are never really used to represent array values; however, they are very convenient during desugaring (and, in particular, in the pattern matching compiler) to treat array pattern just like yet another constructor pattern

isPArrTyCon :: TyCon -> Bool

Check whether a type constructor is the constructor for parallel arrays

isPArrFakeCon :: DataCon -> Bool

Checks whether a data constructor is a fake constructor for parallel arrays

parrTyCon_RDR :: RdrName

parrTyConName :: Name

Equality predicates

eqTyCon_RDR :: RdrName

eqTyCon :: TyCon

eqTyConName :: Name

eqBoxDataCon :: DataCon

Implicit parameter predicates

mkIPName :: FastString -> Unique -> Unique -> Unique -> Unique -> IPName Name