Var

Description

GHC uses several kinds of name internally:

OccName: see OccName

RdrName: see RdrName

Name: see Name

Id: see Id

Var is a synonym for the Id type but it may additionally potentially contain type variables, which have a Kind rather than a Type and only contain some extra details during typechecking.

These Var names may either be global or local, see Var

Global Ids and Vars are those that are imported or correspond to a data constructor, primitive operation, or record selectors. Local Ids and Vars are those bound within an expression (e.g. by a lambda) or at the top level of the module being compiled.

The main data type and synonyms

data Var

Essentially a typed Name, that may also contain some additional information about the Var and it's use sites.

type TyVar = Var

type CoVar = Id

type Id = Var

type KindVar = Var

type DictId = EvId

type DFunId = Id

type EvVar = EvId

type EqVar = EvId

type EvId = Id

type IpId = EvId

Taking Vars apart

varName :: Var -> Name

varUnique :: Var -> Unique

varType :: Var -> Kind

The type or kind of the Var in question

Modifying Vars

setVarName :: Var -> Name -> Var

setVarUnique :: Var -> Unique -> Var

setVarType :: Id -> Type -> Id

Constructing, taking apart, modifying Ids

mkGlobalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

mkLocalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

mkExportedLocalVar :: IdDetails -> Name -> Type -> IdInfo -> Id

Exported Vars will not be removed as dead code

mkCoVar :: Name -> Type -> CoVar

idInfo :: Id -> IdInfo

idDetails :: Id -> IdDetails

lazySetIdInfo :: Id -> IdInfo -> Var

setIdDetails :: Id -> IdDetails -> Id

globaliseId :: Id -> Id

If it's a local, make it global

setIdExported :: Id -> Id

Exports the given local Id. Can also be called on global Ids, such as data constructors and class operations, which are born as global Ids and automatically exported

setIdNotExported :: Id -> Id

We can only do this to LocalIds

Predicates

isId :: Var -> Bool

isTyVar :: Var -> Bool

isTcTyVar :: Var -> Bool

isLocalVar :: Var -> Bool

isLocalVar returns True for type variables as well as local Ids These are the variables that we need to pay attention to when finding free variables, or doing dependency analysis.

isLocalId :: Var -> Bool

isGlobalId :: Var -> Bool

isExportedId :: Var -> Bool

isExportedIdVar means "don't throw this away"

mustHaveLocalBinding :: Var -> Bool

mustHaveLocalBinding returns True of Ids and TyVars that must have a binding in this module. The converse is not quite right: there are some global Ids that must have bindings, such as record selectors. But that doesn't matter, because it's only used for assertions

Constructing TyVars

mkTyVar :: Name -> Kind -> TyVar

mkTcTyVar :: Name -> Kind -> TcTyVarDetails -> TyVar

mkKindVar :: Name -> SuperKind -> KindVar

Taking TyVars apart

tyVarName :: TyVar -> Name

tyVarKind :: TyVar -> Kind

tcTyVarDetails :: TyVar -> TcTyVarDetails

setTcTyVarDetails :: TyVar -> TcTyVarDetails -> TyVar

Modifying TyVars

setTyVarName :: TyVar -> Name -> TyVar

setTyVarUnique :: TyVar -> Unique -> TyVar

setTyVarKind :: TyVar -> Kind -> TyVar

updateTyVarKind :: (Kind -> Kind) -> TyVar -> TyVar

updateTyVarKindM :: Monad m => (Kind -> m Kind) -> TyVar -> m TyVar